Binary variable from a binomial distribution with a logit link

Binary variable from a binomial distribution with a logit link

binary_logit(data)

Arguments

data
vector, scalar, matrix, or data.frame representing the latent variable used to draw the binary outcome

Examples

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = binary_logit(Y1))

#>    ID           Y1 Y2
#> 1  01 -1.400043517  0
#> 2  02  0.255317055  0
#> 3  03 -2.437263611  0
#> 4  04 -0.005571287  0
#> 5  05  0.621552721  0
#> 6  06  1.148411606  1
#> 7  07 -1.821817661  0
#> 8  08 -0.247325302  0
#> 9  09 -0.244199607  1
#> 10 10 -0.282705449  1

binary_logit(runif(10))

#>  [1] 0 1 0 0 1 0 0 0 0 1