Count variable from a binomial distribution with a logit link

Count variable from a binomial distribution with a logit link

binomial_count(data, k)

Arguments

data
vector, scalar, matrix, or data.frame representing the latent variable used to draw the count outcome

k
number of binomial trials, i.e. maximum of the count variable

Examples

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = binomial_count(Y1, 4))

#>    ID          Y1 Y2
#> 1  01 -0.09744510  2
#> 2  02 -0.93584735  1
#> 3  03 -0.01595031  0
#> 4  04 -0.82678895  1
#> 5  05 -1.51239965  0
#> 6  06  0.93536319  4
#> 7  07  0.17648861  3
#> 8  08  0.24368546  2
#> 9  09  1.62354888  3
#> 10 10  0.11203808  2

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = binomial_count(Y1, 2))

#>    ID          Y1 Y2
#> 1  01  0.07003485  0
#> 2  02 -0.63912332  1
#> 3  03 -0.04996490  1
#> 4  04 -0.25148344  0
#> 5  05  0.44479712  1
#> 6  06  2.75541758  2
#> 7  07  0.04653138  1
#> 8  08  0.57770907  1
#> 9  09  0.11819487  1
#> 10 10 -1.91172049  1

binomial_count(runif(50), 3)

#>  [1] 0 1 0 2 2 0 1 3 2 2 1 3 3 1 2 1 1 2 2 1 1 3 1 3 1 3 2 2 1 3 0 3 2 1 1 3 1 1
#> [39] 2 0 2 0 3 3 3 1 3 2 3 1