Creates cross-classified (partially non-nested, joined data) with a fixed correlation structure.

Creates cross-classified (partially non-nested, joined data) with a fixed correlation structure.

cross_level(N = NULL, ID_label = NULL, working_environment_ = NULL,
  by = NULL, ..., data_arguments = quos(...))

Arguments

N
(required) The number of observations in the resulting data frame

ID_label
Internal keyword used to sepcify the name of the ID variable created for the new level. If left empty, this will be the name the level is assigned to as part of a fabricate() call.

working_environment_
Internal keyword not for end user use.

by
The result of a call to join() which specifies how the cross-classified data will be created

...
A variable or series of variables to add to the resulting data frame after the cross-classified data is created.

data_arguments
Internal keyword not for end user use.