Draw a binary variable from a binomial distribution

Drawing binary variables based on probabilities or latent traits is a common task that can be cumbersome.

draw_binary(vector, link = "logit")

Arguments

vector
vector representing either the latent variable used to draw the count outcome (if link is "logit" or "probit") or the probability for the count outcome (if link is "identity")

link
link function between the latent variable and the probability of a postiive outcome, i.e. "logit", "probit", or "identity". For the "identity" link, the latent variable must be a probability.

Examples

fabricate(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1))

#>    ID           Y1 Y2
#> 1  01 -1.400043517  0
#> 2  02  0.255317055  0
#> 3  03 -2.437263611  0
#> 4  04 -0.005571287  0
#> 5  05  0.621552721  0
#> 6  06  1.148411606  1
#> 7  07 -1.821817661  0
#> 8  08 -0.247325302  0
#> 9  09 -0.244199607  1
#> 10 10 -0.282705449  1

fabricate(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1, link = "logit"))

#>    ID          Y1 Y2
#> 1  01 -1.86301149  0
#> 2  02 -0.52201251  0
#> 3  03 -0.05260191  0
#> 4  04  0.54299634  1
#> 5  05 -0.91407483  0
#> 6  06  0.46815442  0
#> 7  07  0.36295126  1
#> 8  08 -1.30454355  0
#> 9  09  0.73777632  1
#> 10 10  1.88850493  1

fabricate(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1, link = "probit"))

#>    ID         Y1 Y2
#> 1  01  0.9353632  1
#> 2  02  0.1764886  0
#> 3  03  0.2436855  1
#> 4  04  1.6235489  1
#> 5  05  0.1120381  1
#> 6  06 -0.1339970  1
#> 7  07 -1.9100875  0
#> 8  08 -0.2792372  0
#> 9  09 -0.3134460  1
#> 10 10  1.0673079  1

fabricate(N = 10, Y1 = runif(N, 0, 1),  Y2 = draw_binary(Y1, link = "identity"))

#>    ID         Y1 Y2
#> 1  01 0.99706914  1
#> 2  02 0.14903547  0
#> 3  03 0.51855664  1
#> 4  04 0.84612005  1
#> 5  05 0.71826972  1
#> 6  06 0.24131402  0
#> 7  07 0.54704337  1
#> 8  08 0.83480182  1
#> 9  09 0.02795603  0
#> 10 10 0.46938430  1

draw_binary(rnorm(10))

#>  [1] 1 0 1 1 1 1 1 1 0 1

draw_binary(rnorm(10), link = "logit")

#>  [1] 1 0 0 0 1 0 0 0 1 1

draw_binary(rnorm(10), link = "probit")

#>  [1] 0 0 0 0 1 1 0 1 0 0

draw_binary(runif(10, 0, 1), link = "identity")

#>  [1] 0 1 0 0 0 0 0 0 0 1