Draw a count variable from a binomial distribution

Drawing count variables based on probabilities or latent traits is a common task that can be cumbersome.

draw_count(vector, k, link = "logit")

Arguments

vector
vector representing either the latent variable used to draw the count outcome (if link is "logit" or "probit") or the probability for the count outcome (if link is "identity")

k
number of binomial trials, i.e. maximum of the count variable

link
link function between the latent variable and the probability of a postiive outcome, i.e. "logit", "probit", or "identity". For the "identity" link, the latent variable must be a probability.

Examples

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = draw_count(Y1, k = 4))

#>    ID         Y1 Y2
#> 1  01 -0.2088827  4
#> 2  02 -1.3994105  0
#> 3  03  0.2585373  2
#> 4  04 -0.4417995  1
#> 5  05  0.5685999  1
#> 6  06  2.1268505  3
#> 7  07  0.4248584  2
#> 8  08 -1.6842815  1
#> 9  09  0.2494018  2
#> 10 10  1.0728383  4

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = draw_count(Y1, k = 4, link = "probit"))

#>    ID          Y1 Y2
#> 1  01  0.02229473  2
#> 2  02  0.60361101  1
#> 3  03 -0.26265057  1
#> 4  04 -0.52826408  3
#> 5  05  0.19214942  3
#> 6  06 -1.14619967  1
#> 7  07  0.84618466  3
#> 8  08  0.08171963  4
#> 9  09 -1.30511701  0
#> 10 10 -0.94491206  1

fabricate_data(N = 10, Y1 = runif(N, 0, 1),  Y2 = draw_count(Y1, k = 4, link = "identity"))

#>    ID         Y1 Y2
#> 1  01 0.43538664  3
#> 2  02 0.55247234  4
#> 3  03 0.20403065  0
#> 4  04 0.03102602  0
#> 5  05 0.96970706  4
#> 6  06 0.17861309  0
#> 7  07 0.77829279  4
#> 8  08 0.88571080  4
#> 9  09 0.83644625  2
#> 10 10 0.60536844  3

draw_count(rnorm(10), k = 4)

#>  [1] 4 1 2 3 3 4 2 2 4 2

draw_count(rnorm(10), k = 4, link = "logit")

#>  [1] 2 1 3 1 4 3 1 4 1 3

draw_count(rnorm(10), k = 4, link = "probit")

#>  [1] 1 1 0 4 4 3 1 4 2 4

draw_count(runif(10, 0, 1), k = 4, link = "identity")

#>  [1] 3 1 1 4 4 1 4 4 4 0