Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4950854
#> Y_2 0.4950854 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.1994969 -1.0202289
#> 2  0.1242529  1.5516487
#> 3  1.2011557  0.5812092
#> 4 -1.8102581 -1.9421005
#> 5 -0.3964831 -0.1289314

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.189976528 -1.333237881
#> 2   002  1.412855016 -1.006965329
#> 3   003 -0.242150450 -1.405944485
#> 4   004  1.592383423  1.695810075
#> 5   005 -0.839994120 -0.966137446
#> 6   006 -1.084186460 -1.225929871
#> 7   007  1.073077998  1.983515034
#> 8   008 -0.300545170  1.177072744
#> 9   009 -1.390621008  0.255621595
#> 10  010  0.668076124  0.488083100
#> 11  011  0.700517426  1.677373812
#> 12  012  0.608224466  0.421621731
#> 13  013  0.717929997  0.683831603
#> 14  014 -1.549208297 -0.060476094
#> 15  015 -0.744157272 -0.933996571
#> 16  016  0.056664174 -1.032645443
#> 17  017  1.449934454  1.821023936
#> 18  018  0.412759611 -0.615010503
#> 19  019 -0.050567317 -0.035651469
#> 20  020  0.617411516  0.097112954
#> 21  021  0.319710209  0.267465782
#> 22  022  0.527478441 -0.907601131
#> 23  023 -0.078430010 -0.406409076
#> 24  024  0.482921029 -0.802426591
#> 25  025 -0.766741783 -1.192279022
#> 26  026  0.390070400  0.647546061
#> 27  027 -0.399459645 -0.198140249
#> 28  028  0.043609671  0.525687072
#> 29  029 -0.255051899  0.251584691
#> 30  030 -1.160088723 -1.686780081
#> 31  031  1.001313285  0.436539305
#> 32  032  1.115420772  1.155818807
#> 33  033 -0.528091720 -0.728873099
#> 34  034  0.115791905 -0.788425289
#> 35  035 -0.986079878 -0.819424354
#> 36  036  0.902971345  0.378686562
#> 37  037 -0.337203024 -0.319639121
#> 38  038  0.266228119  1.214745306
#> 39  039  0.670034145  0.940299929
#> 40  040 -0.362432275 -0.523373679
#> 41  041 -0.484109706  1.481089801
#> 42  042 -1.219476791 -0.823928748
#> 43  043  0.319979597  1.193698967
#> 44  044  0.415656240  0.887980395
#> 45  045  1.289900530 -0.195325919
#> 46  046  0.841612320  0.005450021
#> 47  047  0.750488847  0.104219635
#> 48  048 -1.729696044 -1.777133080
#> 49  049 -0.183773109 -0.016013120
#> 50  050  0.398843970  0.299105301
#> 51  051 -1.194425087 -1.108428632
#> 52  052  0.922436560  0.549862529
#> 53  053  0.866882362  1.048214905
#> 54  054  0.803400363 -0.103406048
#> 55  055 -0.402222253  0.188934899
#> 56  056 -0.007543088  0.784970144
#> 57  057  1.752637493  1.829509877
#> 58  058 -1.048212049 -1.116049248
#> 59  059 -0.406763598  0.026524449
#> 60  060  0.869192858  0.918288261
#> 61  061 -1.354411309 -1.387211357
#> 62  062  0.329046647 -0.181878129
#> 63  063  0.870363630  1.226794168
#> 64  064 -0.229292251  0.188653885
#> 65  065 -0.216255545  0.362949715
#> 66  066  0.456064339 -1.019094315
#> 67  067 -0.223759956  1.100045274
#> 68  068 -0.155674592  0.874946874
#> 69  069 -1.178357292 -0.860983400
#> 70  070  0.584916399 -0.526987389
#> 71  071 -0.899393204 -0.278182035
#> 72  072 -0.057742485  1.751366066
#> 73  073  0.978017440  2.092003446
#> 74  074  0.008022025  0.473372560
#> 75  075  0.117739926 -0.950889803
#> 76  076 -1.679074123 -1.423133839
#> 77  077  0.848177714  0.070958780
#> 78  078 -0.498916963 -1.449235200
#> 79  079 -0.580421298  0.650095034
#> 80  080 -0.883649279 -1.173970609
#> 81  081  2.570169666  1.324920081
#> 82  082  1.807632842  0.882522219
#> 83  083  0.455624348  0.103866077
#> 84  084  0.040525891 -0.001104937
#> 85  085 -0.580659587  0.148637836
#> 86  086 -2.027985058 -1.734582029
#> 87  087 -2.953651567  0.633293909
#> 88  088 -1.226997074 -0.471254068
#> 89  089 -1.752657261 -1.352529957
#> 90  090  0.038096318  0.867442031
#> 91  091 -1.225562767 -1.530699010
#> 92  092  0.343192694  1.053296016
#> 93  093 -0.811506397 -0.372892580
#> 94  094  0.110476704  0.468700401
#> 95  095  0.878786342  0.754548691
#> 96  096 -0.431418505 -1.124634591
#> 97  097  0.377359091  0.951705619
#> 98  098  0.188547397 -0.028612118
#> 99  099 -1.823368274  1.214212889
#> 100 100  0.895749519 -0.695917612

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.09190104 -0.074962942
#> 2   002  0.99251365  0.269411464
#> 3   003 -0.27053963 -1.724024670
#> 4   004  0.07380717 -2.067008310
#> 5   005 -0.18180776 -0.046953332
#> 6   006 -0.83419979 -0.172202018
#> 7   007  1.19146475  1.686077917
#> 8   008  1.77479531  0.248151658
#> 9   009  0.72184837  0.611632366
#> 10  010 -1.37563307  1.093108208
#> 11  011  1.38618390  0.822640469
#> 12  012  0.87958030 -0.566273122
#> 13  013  1.51511155  1.038555779
#> 14  014 -1.82143725 -0.806836786
#> 15  015  0.32262846  0.204608424
#> 16  016  0.27023041 -0.084091435
#> 17  017 -1.20432940 -0.917073561
#> 18  018  0.37942339  0.397436004
#> 19  019  0.88921434 -0.009209638
#> 20  020 -0.41907520  1.025661290
#> 21  021  0.92553188  0.387732730
#> 22  022  0.49313166  0.010771665
#> 23  023  0.38760679 -0.348012819
#> 24  024 -0.17263959 -0.187874849
#> 25  025 -0.08576395 -0.370824783
#> 26  026 -1.33286401 -2.103760001
#> 27  027  0.64605328 -1.265469668
#> 28  028 -0.48238288 -0.304483220
#> 29  029  0.14965950  0.345980181
#> 30  030 -0.27545878 -0.225414620
#> 31  031  0.96595958 -0.079314854
#> 32  032 -0.88804058 -1.024722966
#> 33  033 -0.80641912  0.421081485
#> 34  034 -1.16066748 -1.354532674
#> 35  035  1.34233843 -0.650732674
#> 36  036  1.45213178  1.022226110
#> 37  037 -0.09158819 -0.743504512
#> 38  038  1.52081104  0.198055719
#> 39  039 -0.70229315 -1.313295504
#> 40  040  1.62038694  1.186902169
#> 41  041 -0.20361864  0.017512981
#> 42  042 -0.87693436 -0.462299059
#> 43  043 -1.13939597 -0.135876333
#> 44  044  2.01944383 -0.606598412
#> 45  045  2.34465299  1.471118063
#> 46  046 -0.19428801 -0.967998843
#> 47  047 -0.66163658 -0.068578032
#> 48  048 -0.90033217 -0.446768829
#> 49  049  0.73112332  0.241496606
#> 50  050 -1.59170482 -2.257790322
#> 51  051 -0.05741879 -1.147073019
#> 52  052  0.18828947 -0.852617482
#> 53  053  1.53196796  2.033574643
#> 54  054 -0.98022287 -0.938869101
#> 55  055 -0.80132642 -0.854408103
#> 56  056  0.21310794 -1.314291140
#> 57  057  0.74581569  0.582247172
#> 58  058  1.66247063 -0.472624292
#> 59  059  1.55921604  1.422382452
#> 60  060  2.18936576 -0.714854881
#> 61  061  0.20380517  0.540817814
#> 62  062 -1.29615266 -0.531642001
#> 63  063  0.29495501 -0.374423181
#> 64  064 -0.03372819 -0.194697065
#> 65  065 -1.03029407  0.476966958
#> 66  066  0.47100189  0.349308047
#> 67  067  0.82536697  0.581952195
#> 68  068 -0.13184580  0.576311392
#> 69  069 -1.62086597 -2.100905634
#> 70  070  0.75592717  0.439375903
#> 71  071 -1.79021701 -1.279745763
#> 72  072  0.60623669  0.339972339
#> 73  073 -0.30180849  0.334623610
#> 74  074 -0.14141734 -1.155372935
#> 75  075 -1.15401757 -0.221786822
#> 76  076  0.58588109 -0.064907240
#> 77  077 -0.17074707  0.265431108
#> 78  078  0.39747636  0.233386960
#> 79  079 -1.64340947 -0.374753088
#> 80  080  0.01793171  0.433364216
#> 81  081 -2.06923196 -2.117336241
#> 82  082  0.89599034  1.094283291
#> 83  083  1.23927193  0.850584627
#> 84  084 -0.14799136 -0.600223646
#> 85  085  0.16225136  0.022829937
#> 86  086  1.08365667 -0.241540985
#> 87  087 -1.72236913 -0.968299613
#> 88  088 -1.61695524 -0.259102407
#> 89  089  0.16351291 -1.057271955
#> 90  090  0.10402438 -0.463894278
#> 91  091  1.38015354  0.191303966
#> 92  092  1.20885272  0.196576831
#> 93  093 -0.64213279 -1.074664281
#> 94  094 -0.56629456 -0.220588171
#> 95  095 -0.36889273 -0.177965137
#> 96  096 -1.51525430 -0.227765726
#> 97  097 -0.77364467 -0.987148928
#> 98  098  1.46122042  2.054781722
#> 99  099  0.31238758 -1.503412290
#> 100 100 -1.14490539 -1.605429961