Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms – standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.20837615 -1.61642098  0.01102375  0.06771514 -0.79873109  1.09170191
#>  [7] -1.50790361 -1.81571661 -1.73732608  1.65692101  0.67939628 -0.56389767
#> [13] -0.04217591 -0.37582103  1.77957414  1.55054037 -1.57378193 -2.36118445
#> [19]  0.58769817  0.80938753  1.88568428 -1.12310312 -0.77846739  1.66999219
#> [25]  0.46376166  0.92701170 -0.62857207  0.98836310 -0.59084932  1.08455683
#> [31]  0.87602961 -1.08058824 -2.37241944  1.02064030  2.13826841 -0.46833413
#> [37] -1.11087264  0.08384998 -0.18172657  3.61670587  0.65658622 -1.14374495
#> [43] -3.06085751  0.27123452  0.89229635 -0.12883584 -0.42647205 -0.67488526
#> [49]  0.35458870  0.75983842

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  5.883550 10.178620 10.044252  9.915881 10.384621  9.356091 15.075908
#>  [8] 11.475382  9.555016  9.741351 11.410901  8.064338 11.775417 11.325256
#> [15] 10.392889 13.001148  9.995516  8.681418  8.997935  7.067731 10.749302
#> [22]  9.403433 12.186681  7.632452  6.645163 13.386815  8.151098  6.158471
#> [29]  5.395236  6.231912  8.113161 12.501242 16.529585  6.759916  8.347730
#> [36]  7.608051 16.600535 11.109468 10.998363  8.699683  9.181862  9.026431
#> [43] 10.393688  5.254076  8.903343 14.507954  9.583014 12.092896  3.432835
#> [50] 11.008512

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.6617668 -13.9738189   4.3331829  -5.2655252   1.3345065  -1.4290741
#>  [7]  -4.0988554   1.0732485  11.1382780  -2.2714261 -19.3404232  -0.6864183
#> [13]  -1.5447276 -10.5772020  21.1219793 -13.2730263  -9.6862693   1.3253769
#> [19]   8.0493048  -5.1338750  -6.8937865  -2.7991250   7.3559111   9.0407351
#> [25]  16.3863801  -1.7553888   6.2579463  -2.6370452   4.4479243  -1.9099628
#> [31]  -0.7353908   3.4137752  -5.1677517  -1.3170232   9.5923464 -15.5906743
#> [37]  -4.0671215   6.6865011  13.3270315   5.3585906  -3.3193311   3.4256877
#> [43]   3.8342999  -0.1225453   2.9502685 -12.8693451   6.4198411  -4.5249828
#> [49]  -2.8601113  11.9403624

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3267587