Fabricate data

Fabricate data

fabricate_data(..., N = NULL, ID_label = NULL, data = NULL)

Arguments

...
Data generating arguments, such as my_var = rnorm(N). You may also pass level() arguments, which define a level of a multi-level dataset. For example, you could send to ... level(my_level, var = rnorm). See examples.

N
number of units to draw

ID_label
variable name for ID variable, i.e. citizen_ID (optional)

data
user-provided data that forms the basis of the fabrication, i.e. you can add variables to existing data. Provide either N or data (N is the number of rows of the data if data is provided).

Value

data.frame

Examples

# Draw a single-level dataset with no covariates
df <- fabricate_data(N = 100)
head(df)

#>    ID
#> 1 001
#> 2 002
#> 3 003
#> 4 004
#> 5 005
#> 6 006

# Draw a single-level dataset with a covariate
df <- fabricate_data(
  N = 100,
  height_ft = runif(N, 3.5, 8)
)
head(df)

#>    ID height_ft
#> 1 001  6.879562
#> 2 002  6.735563
#> 3 003  5.819099
#> 4 004  7.886763
#> 5 005  5.123499
#> 6 006  4.266780

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate_data(
 regions = level(N = 5),
 cities = level(N = 10, pollution = rnorm(N, mean = 5)))
head(df)

#>   regions cities pollution
#> 1       1     01  5.065487
#> 2       1     02  3.901491
#> 3       1     03  4.366822
#> 4       1     04  2.936346
#> 5       1     05  7.648932
#> 6       1     06  3.846602