Fabricate a Level of Data for Multi-Level Hierarchical Data

Fabricate a Level of Data for Multi-Level Hierarchical Data

level(ID_label, N = NULL, ..., level_data = NULL, by = NULL,
  data = NULL)

Arguments

ID_label
variable name for ID variable, i.e. citizen_ID (optional)

N
number of units to draw in the level

...
Data generating arguments, such as my_var = rnorm(N). You may also pass level() arguments, which define a level of a multi-level dataset. For example, you could send to ... level(my_level, var = rnorm). See examples.

level_data
user-provided data that forms the basis of the fabrication at this level, i.e. you can add variables to existing data. Provide either N or data (N is the number of rows of the data if data is provided).

by
name of variable to merge by to level above, if the level is not at the top of the hierarchy.

data
system option that is used by fabricate_data to send data from upper levels of hierarchy to this level. Do not provide data.

Examples

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate_data(
 regions = level(N = 5),
 cities = level(N = 10, pollution = rnorm(N, mean = 5)))
head(df)

#>   regions cities pollution
#> 1       1     01  4.171396
#> 2       1     02  5.289774
#> 3       1     03  4.519947
#> 4       1     04  4.395171
#> 5       1     05  6.460110
#> 6       1     06  5.149679

# Use existing data in a level using level_data
region_data <- data.frame(capital = c(1, 0, 0, 0, 0))
fabricate_data(regions = level(level_data = region_data,
                               gdp = runif(5)),
               cities = level(N = 5,
                              subways = rnorm(N, mean = 5)))

#>    capital regions       gdp cities  subways
#> 1        1       1 0.3444290     01 3.789521
#> 2        1       1 0.3444290     02 5.380149
#> 3        1       1 0.3444290     03 4.932614
#> 4        1       1 0.3444290     04 5.280100
#> 5        1       1 0.3444290     05 5.330308
#> 6        0       2 0.7608582     06 5.284752
#> 7        0       2 0.7608582     07 5.428594
#> 8        0       2 0.7608582     08 4.879033
#> 9        0       2 0.7608582     09 5.458242
#> 10       0       2 0.7608582     10 5.616535
#> 11       0       3 0.7311036     11 4.824951
#> 12       0       3 0.7311036     12 4.877169
#> 13       0       3 0.7311036     13 3.527308
#> 14       0       3 0.7311036     14 5.123693
#> 15       0       3 0.7311036     15 3.904408
#> 16       0       4 0.8524324     16 5.388342
#> 17       0       4 0.8524324     17 4.244970
#> 18       0       4 0.8524324     18 3.768397
#> 19       0       4 0.8524324     19 5.764266
#> 20       0       4 0.8524324     20 4.354195
#> 21       0       5 0.9758180     21 3.958019
#> 22       0       5 0.9758180     22 4.162794
#> 23       0       5 0.9758180     23 6.026353
#> 24       0       5 0.9758180     24 5.853874
#> 25       0       5 0.9758180     25 5.825224