Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0        Y_Z_1
#> 1  01  0.3969364  0.3969364  0.496936358
#> 2  02  0.7274949  0.7274949  0.827494886
#> 3  03 -0.3641801 -0.3641801 -0.264180092
#> 4  04  0.4284630  0.4284630  0.528462996
#> 5  05  1.3634928  1.3634928  1.463492823
#> 6  06  1.2300580  1.2300580  1.330057953
#> 7  07 -0.1030017 -0.1030017 -0.003001675
#> 8  08 -1.0911044 -1.0911044 -0.991104421
#> 9  09  2.0468823  2.0468823  2.146882306
#> 10 10 -0.1250526 -0.1250526 -0.025052642

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01 -0.41190068 -0.41190068 -0.311900681
#> 2  02 -1.16465475 -1.16465475 -1.064654747
#> 3  03 -0.38878636 -0.38878636 -0.288786361
#> 4  04 -0.02959099 -0.02959099  0.070409011
#> 5  05 -0.10951299 -0.10951299 -0.009512988
#> 6  06 -0.80882159 -0.80882159 -0.708821589
#> 7  07  0.91885320  0.91885320  1.018853199
#> 8  08 -0.17512524 -0.17512524 -0.075125245
#> 9  09  0.10844145  0.10844145  0.208441445
#> 10 10  0.65289296  0.65289296  0.752892959

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U        Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01  1.864609876  1.964609876  2.06460988  2.164609876
#> 2  02 -0.727595820 -0.627595820 -0.52759582 -0.427595820
#> 3  03 -0.309767047 -0.209767047 -0.10976705 -0.009767047
#> 4  04 -0.001512235  0.098487765  0.19848776  0.298487765
#> 5  05  0.072130396  0.172130396  0.27213040  0.372130396
#> 6  06  0.320936649  0.420936649  0.52093665  0.620936649
#> 7  07 -0.545614399 -0.445614399 -0.34561440 -0.245614399
#> 8  08  2.158878742  2.258878742  2.35887874  2.458878742
#> 9  09  1.026827359  1.126827359  1.22682736  1.326827359
#> 10 10 -0.102969928 -0.002969928  0.09703007  0.197030072

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.3421831   0.3421831   0.4421831   0.6421831  1.24218308
#> 2  02 -0.8846198  -0.8846198  -0.7846198  -0.5846198  0.01538016
#> 3  03 -2.5457334  -2.5457334  -2.4457334  -2.2457334 -1.64573340
#> 4  04 -0.9430532  -0.9430532  -0.8430532  -0.6430532 -0.04305323
#> 5  05 -0.4709147  -0.4709147  -0.3709147  -0.1709147  0.42908529
#> 6  06 -1.1368760  -1.1368760  -1.0368760  -0.8368760 -0.23687597
#> 7  07 -0.6174131  -0.6174131  -0.5174131  -0.3174131  0.28258689
#> 8  08  0.7112748   0.7112748   0.8112748   1.0112748  1.61127485
#> 9  09  0.2794528   0.2794528   0.3794528   0.5794528  1.17945275
#> 10 10 -0.6181454  -0.6181454  -0.5181454  -0.3181454  0.28185458