Creates cross-classified (partially non-nested, joined data) with a fixed correlation structure.

Creates cross-classified (partially non-nested, joined data) with a fixed correlation structure.

cross_level(N = NULL, ID_label = NULL, working_environment_ = NULL,
  by = NULL, ..., data_arguments = quos(...))

Arguments

N
(required) The number of observations in the resulting data frame

ID_label
Internal keyword used to sepcify the name of the ID variable created for the new level. If left empty, this will be the name the level is assigned to as part of a fabricate() call.

working_environment_
Internal keyword not for end user use.

by
The result of a call to join() which specifies how the cross-classified data will be created

...
A variable or series of variables to add to the resulting data frame after the cross-classified data is created.

data_arguments
Internal keyword not for end user use.

Value

data.frame

Examples

# Generate cross-classified data and merge, no correlation
students <- fabricate(
 primary_school = add_level(N = 20, ps_quality = runif(N, 1, 10)),
 secondary_school = add_level(N = 15, ss_quality = runif(N, 1, 10), nest=FALSE),
 students = cross_level(N = 500, by = join(primary_school, secondary_school))
)
head(students)

#>   primary_school ps_quality secondary_school ss_quality students
#> 1             09   7.595938               01   3.609031      001
#> 2             15   4.620954               07   1.463016      002
#> 3             19   4.498312               10   7.197004      003
#> 4             04   2.414876               09   7.262415      004
#> 5             05   1.066595               12   3.030063      005
#> 6             14   3.883472               08   5.771912      006

# Cross-classified data with a correlation structure
students <- fabricate(
 primary_school = add_level(N = 20, ps_quality = runif(N, 1, 10)),
 secondary_school = add_level(N = 15, ss_quality = runif(N, 1, 10), nest=FALSE),
 students = cross_level(N = 500, by = join(ps_quality, ss_quality, rho = 0.5))
)
cor(students$ps_quality, students$ss_quality)

#> [1] 0.4690153