Draw discrete variables including binary, binomial count, poisson count, ordered, and categorical

Drawing discrete data based on probabilities or latent traits is a common task that can be cumbersome. Each function in our discrete drawing set creates a different type of discrete data: draw_binary creates binary 0/1 data, draw_binomial creates binomial data (repeated trial binary data), draw_categorical creates categorical data, draw_ordered transforms latent data into observed ordered categories, draw_count creates count data (poisson-distributed). draw_likert is an alias to draw_ordered that pre-specifies break labels and offers default breaks appropriate for a likert survey question.

draw_binomial(prob, trials = 1, N = length(prob), link = "identity")

draw_categorical(prob, N = NULL, link = "identity")

draw_ordered(x, breaks = c(-1, 0, 1), break_labels = NULL, N = length(x),
  link = "identity")

draw_count(mean, N = length(mean), link = "identity")

draw_binary(prob, N = length(prob), link = "identity")

draw_likert(x, type = 7, breaks = NULL, N = length(x),
  link = "identity")

Arguments

prob
A number or vector of numbers representing the probability for binary or binomial outcomes; or a number, vector, or matrix of numbers representing probabilities for categorical outcomes. If you supply a link function, these underlying probabilities will be transformed.

trials
for `draw_binomial`, the number of trials for each observation

N
number of units to draw. Defaults to the length of the vector of probabilities or latent data you provided.

link
link function between the latent variable and the probability of a postiive outcome, e.g. "logit", "probit", or "identity". For the "identity" link, the latent variable must be a probability.

x
for `draw_ordered` or `draw_likert`, the latent data for each observation.

breaks
vector of breaks to cut a latent outcome into ordered categories with `draw_ordered` or `draw_likert`

break_labels
vector of labels for the breaks to cut a latent outcome into ordered categories with `draw_ordered`.

mean
for `draw_count`, the mean number of count units for each observation

type
Type of Likert scale data for `draw_likert`. Valid options are 4, 5, and 7. Type corresponds to the number of categories in the Likert scale.

Details

For variables with intra-cluster correlations, see draw_binary_icc and draw_normal_icc

Examples

# Drawing binary values (success or failure, treatment assignment)
fabricate(N = 3,
   p = c(0, .5, 1),
   binary = draw_binary(prob = p))

#>   ID   p binary
#> 1  1 0.0      0
#> 2  2 0.5      0
#> 3  3 1.0      1

# Drawing binary values with probit link (transforming continuous data
# into a probability range).
fabricate(N = 3,
   x = 10 * rnorm(N),
   binary = draw_binary(prob = x, link = "probit"))

#>   ID     x binary
#> 1  1   3.3      1
#> 2  2  -5.7      0
#> 3  3 -10.5      0

# Repeated trials: `draw_binomial`
fabricate(N = 3,
   p = c(0, .5, 1),
   binomial = draw_binomial(prob = p, trials = 10))

#>   ID   p binomial
#> 1  1 0.0        0
#> 2  2 0.5        5
#> 3  3 1.0       10

# Ordered data: transforming latent data into observed, ordinal data.
# useful for survey responses.
fabricate(N = 3,
   x = 5 * rnorm(N),
   ordered = draw_ordered(x = x,
                          breaks = c(-Inf, -1, 1, Inf)))

#>   ID     x ordered
#> 1  1  0.94       2
#> 2  2  6.25       3
#> 3  3 -2.74       1

# Providing break labels for latent data.
fabricate(N = 3,
   x = 5 * rnorm(N),
   ordered = draw_ordered(x = x,
                          breaks = c(-Inf, -1, 1, Inf),
                          break_labels = c("Not at all concerned",
                                           "Somewhat concerned",
                                           "Very concerned")))

#>   ID    x              ordered
#> 1  1 -8.6 Not at all concerned
#> 2  2 -3.5 Not at all concerned
#> 3  3  9.5       Very concerned

# Likert data: often used for survey data
fabricate(N = 10,
          support_free_college = draw_likert(x = rnorm(N),
                                             type = 5))

#>    ID support_free_college
#> 1  01                Agree
#> 2  02                Agree
#> 3  03 Don't Know / Neutral
#> 4  04                Agree
#> 5  05             Disagree
#> 6  06             Disagree
#> 7  07 Don't Know / Neutral
#> 8  08 Don't Know / Neutral
#> 9  09       Strongly Agree
#> 10 10                Agree

# Count data: useful for rates of occurrences over time.
fabricate(N = 5,
   x = c(0, 5, 25, 50, 100),
   theft_rate = draw_count(mean=x))

#>   ID   x theft_rate
#> 1  1   0          0
#> 2  2   5          2
#> 3  3  25         27
#> 4  4  50         55
#> 5  5 100         87

# Categorical data: useful for demographic data.
fabricate(N = 6, p1 = runif(N), p2 = runif(N), p3 = runif(N),
          cat = draw_categorical(cbind(p1, p2, p3)))

#>   ID    p1   p2     p3 cat
#> 1  1 0.571 0.16 0.6591   3
#> 2  2 0.679 0.18 0.6603   3
#> 3  3 0.089 0.20 0.0024   2
#> 4  4 0.785 0.36 0.9934   3
#> 5  5 0.227 0.18 0.6275   1
#> 6  6 0.448 0.56 0.0146   2