Draw multivariate random variables

Draw multivariate random variables

draw_multivariate(formula, sep = "_")

Arguments

formula
Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep
Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(10, 3, 3, 2), 2, 2)
))

#>          Y_1        Y_2
#> 1  0.9176093  1.8645566
#> 2  1.9578052  2.2880898
#> 3  5.9679291  1.1697457
#> 4  3.8747084  0.1653347
#> 5 -3.3949159 -0.9381202

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(10, 3, 3, 2), 2, 2)
))

#>           Y_1        Y_2
#> 1 -5.15030297 -0.8920808
#> 2 -0.04880929  2.3829329
#> 3 -2.96783075 -0.8560439
#> 4 -0.83464494 -1.8046956
#> 5  3.57069613  2.7110318

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(10, 3, 3, 2), 2, 2)
  ))
)

#>      ID         Y_1          Y_2
#> 1   001 -4.86946632 -0.779589983
#> 2   002  1.74759450  2.553011640
#> 3   003  0.71986642  1.031127284
#> 4   004 -1.24928742  0.206474613
#> 5   005  1.67678829  1.341170745
#> 6   006  1.14926646 -0.347937098
#> 7   007  0.40450707  0.197516222
#> 8   008 -3.15307454  0.913693096
#> 9   009  0.45569308  1.495348054
#> 10  010 -7.35810884 -0.574551604
#> 11  011 -3.59590342 -0.662529968
#> 12  012 -2.88825495  0.868078818
#> 13  013  3.81324204  1.228409641
#> 14  014 -1.07072441 -0.226011496
#> 15  015  1.48694162  1.141546036
#> 16  016 -3.98903810 -1.499032489
#> 17  017 -0.30048684  0.598026075
#> 18  018  1.13907944  0.732657789
#> 19  019  2.38336571  0.137743364
#> 20  020  5.01986591  0.589476689
#> 21  021 -4.90303444 -0.438394340
#> 22  022  1.38181746 -0.505903295
#> 23  023  0.51053342 -1.109358569
#> 24  024 -4.13934725 -0.656746122
#> 25  025  3.88990600  2.450411514
#> 26  026  1.92112154  0.255762777
#> 27  027  1.90248593  1.696705279
#> 28  028  6.01969223  1.421546468
#> 29  029  0.91497063 -0.154663631
#> 30  030  1.08913978  0.251973385
#> 31  031  0.70440705  0.502265531
#> 32  032 -1.75464763  0.444429594
#> 33  033  1.57225631  0.108543425
#> 34  034  0.13551280 -1.024630790
#> 35  035  2.53618836  0.179795901
#> 36  036  6.65466273  2.727767410
#> 37  037  0.25012253  1.692271414
#> 38  038 -2.55776741 -1.502797993
#> 39  039  3.45176208 -0.145194744
#> 40  040  1.66311136  0.388400391
#> 41  041  3.35041152 -0.365610849
#> 42  042  4.82758024  1.190928521
#> 43  043  0.62531170 -0.135642784
#> 44  044 -5.43988153 -0.275010928
#> 45  045 -3.44513094 -0.305845517
#> 46  046  3.75594099  0.756543341
#> 47  047  2.49905438  2.045414849
#> 48  048 -9.86870135 -3.003279042
#> 49  049  3.50991948  0.687805696
#> 50  050  2.19748060  0.750368768
#> 51  051  0.45312377  0.721931840
#> 52  052 -2.06278957 -1.610784924
#> 53  053  0.39209762 -0.650212777
#> 54  054 -4.18382797 -1.220881812
#> 55  055 -0.50811097  0.622215537
#> 56  056  1.58543185  1.849641359
#> 57  057 -2.38881735  0.296016476
#> 58  058 -0.07032967  0.007587753
#> 59  059 -2.27770714 -0.469419526
#> 60  060 -6.16724833 -2.567460222
#> 61  061  0.46925316 -0.810166213
#> 62  062 -1.06804283  0.682629408
#> 63  063 -0.40267981  1.437625405
#> 64  064  2.87325474  0.093825281
#> 65  065 -4.09231576 -2.432327427
#> 66  066  2.61078022 -0.240337377
#> 67  067 -4.20916110 -1.327761304
#> 68  068  2.85398918 -0.052635042
#> 69  069 -5.76939258 -1.518319770
#> 70  070 -2.29252661 -0.323817839
#> 71  071 -0.42993317  0.352395121
#> 72  072 -1.55564755  1.263057271
#> 73  073 -0.52338985  0.822178754
#> 74  074 -0.84173950  0.080986519
#> 75  075 -6.13663222 -2.423396629
#> 76  076  3.00960089  0.952404777
#> 77  077  2.11511888 -0.486266557
#> 78  078 -0.97175366  0.232255405
#> 79  079  0.77488716 -0.247221059
#> 80  080 -0.90089142 -1.794504217
#> 81  081 -0.24799988  0.001756459
#> 82  082  4.71031544  1.985614732
#> 83  083 -1.26232742 -0.974445972
#> 84  084 -3.00292764 -0.161986622
#> 85  085  4.16074223  1.128559736
#> 86  086  3.14844229  1.870955999
#> 87  087  3.75885342  0.625846319
#> 88  088  0.27950783 -1.075704359
#> 89  089 -0.23066682  0.917087692
#> 90  090  1.09860523 -0.372025640
#> 91  091 -0.05813029  0.428269807
#> 92  092 -2.95735763  0.219186958
#> 93  093  2.43421529  0.565337496
#> 94  094 -1.74698870 -0.382856701
#> 95  095  4.42242321  3.227530489
#> 96  096  3.70951602 -0.782810035
#> 97  097 -6.49930053 -1.305351167
#> 98  098  5.94654722  1.577779278
#> 99  099  2.71754944  1.389290652
#> 100 100  5.91373720  2.725663966

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(10, 3, 3, 2), 2, 2)
))

#>      ID        Y.1          Y.2
#> 1   001 -1.0354242 -1.398870466
#> 2   002  3.5740501  1.547368700
#> 3   003 -4.1870600 -1.304999270
#> 4   004 -4.8414452  0.197977737
#> 5   005  0.4919233  1.143354554
#> 6   006  1.5936615 -0.326469366
#> 7   007 -3.2896921 -1.651452633
#> 8   008 -1.7910010 -1.663697096
#> 9   009  0.1655088 -0.821181656
#> 10  010 -7.3210720 -2.877947132
#> 11  011 -8.3451002 -2.084673923
#> 12  012 -3.6408036 -1.792099093
#> 13  013  3.7662359  1.823955541
#> 14  014 -1.3135833 -0.257275637
#> 15  015 -3.1834517 -0.252941319
#> 16  016  1.2954051  1.071187742
#> 17  017 -4.4599088 -3.223056433
#> 18  018 -1.6148203 -1.424838358
#> 19  019 -1.3595606 -0.809654705
#> 20  020  2.2140625  1.164805071
#> 21  021 -2.7329891 -1.859689185
#> 22  022 -0.1037395 -1.289652690
#> 23  023  2.4057289  0.852020232
#> 24  024 -4.1343785 -2.207618412
#> 25  025 -2.5215859 -2.070321661
#> 26  026 -3.2830016 -0.807806337
#> 27  027 -1.7813931 -0.263124363
#> 28  028  3.0169647  1.307441743
#> 29  029 -3.9913713 -0.483556688
#> 30  030  1.5985075  0.632506032
#> 31  031  0.7932717  0.342475771
#> 32  032  1.2581732 -0.333748248
#> 33  033 -0.1916102 -0.195334776
#> 34  034  1.1899019 -0.970096465
#> 35  035  3.9703231  2.498219715
#> 36  036 -0.1390226  0.841267064
#> 37  037 -1.8020387  0.984985757
#> 38  038 -0.7241288  0.058013907
#> 39  039  5.3536461  1.739360758
#> 40  040 -2.0237148 -0.665225587
#> 41  041  2.1988662  3.841562643
#> 42  042 -3.0261318 -1.030170082
#> 43  043  5.1794678  1.830261506
#> 44  044 -3.1830931 -1.682006550
#> 45  045  0.2084089  0.518251270
#> 46  046 -5.7749537 -2.593973872
#> 47  047  3.3946380  1.741364806
#> 48  048  0.8266407  0.453803486
#> 49  049  2.2063968  2.758514297
#> 50  050 -0.9152999  1.312158790
#> 51  051  2.0046755  1.843286671
#> 52  052 -1.1252029 -2.059308989
#> 53  053  0.4028508  0.808697363
#> 54  054  0.4047831 -1.516509146
#> 55  055 -1.5865432  0.998696527
#> 56  056 -0.1590648  0.781296835
#> 57  057 -3.7686144 -0.724435429
#> 58  058 -1.0484014 -0.773029660
#> 59  059 -0.7101805 -1.260668631
#> 60  060  2.4276600  1.420238471
#> 61  061 -3.8802879 -2.984383267
#> 62  062  0.6741368 -0.013484532
#> 63  063 -1.0195118  0.202869804
#> 64  064  3.1428559  1.915967683
#> 65  065  7.4350354  2.125840499
#> 66  066  1.6997933  0.665145695
#> 67  067  0.3245205 -0.212336078
#> 68  068 -3.3987045 -0.630618473
#> 69  069 -0.7532270  0.003289206
#> 70  070  1.8160258 -0.263394577
#> 71  071 -5.2497698 -0.110130557
#> 72  072  1.8528063  0.519043094
#> 73  073 -5.3856422 -1.408415400
#> 74  074  3.1085882  1.172604918
#> 75  075 -3.7181884 -1.669250351
#> 76  076 -3.5739606 -0.575762887
#> 77  077  1.8642672  0.825336492
#> 78  078 -1.6800578 -0.275892138
#> 79  079 -1.4242786 -0.552402115
#> 80  080 -4.5575620 -1.895747276
#> 81  081 -2.6009718 -1.447193082
#> 82  082  1.7008379  0.526466917
#> 83  083 -4.5538927  0.178677094
#> 84  084 -2.1815697  0.379301711
#> 85  085  4.7659527 -0.574382832
#> 86  086 -0.9407620 -0.499567648
#> 87  087 -3.4005293 -2.322850141
#> 88  088 -2.7353076 -1.370301305
#> 89  089  2.6490169  0.304484333
#> 90  090  1.4976918 -0.183208227
#> 91  091 -1.2210803 -1.333008540
#> 92  092 -0.4740386 -0.527786286
#> 93  093  4.1838957  1.342054070
#> 94  094 -3.4244922 -2.022454676
#> 95  095 -0.8174643  0.024476323
#> 96  096 -1.7703013  0.244402995
#> 97  097  1.6848667  0.559879049
#> 98  098 -4.7117331 -0.318811044
#> 99  099  4.2133672  1.022434179
#> 100 100 -2.1712793 -0.144604090