Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4554171
#> Y_2 0.4554171 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1 -0.003090544  1.2345761
#> 2 -0.318068026  0.1907223
#> 3 -0.962135266 -1.6070121
#> 4 -0.192234102  0.1568677
#> 5 -0.965804524 -2.6653879

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.173824870  0.52041316
#> 2   002 -0.190972701  0.04714751
#> 3   003 -1.369807071 -1.32015800
#> 4   004  0.880566546 -0.31886067
#> 5   005  0.962149488  1.16677100
#> 6   006 -0.335933886 -2.15181533
#> 7   007 -0.976825616 -1.21176862
#> 8   008  0.484682173  0.75626008
#> 9   009 -0.101591251  0.51130640
#> 10  010  1.289422946  1.19629590
#> 11  011 -0.721975969 -0.77399641
#> 12  012  1.191095383  0.53501247
#> 13  013  0.505950808  1.04730257
#> 14  014 -0.285551658 -0.06078542
#> 15  015  0.071258727 -0.04643586
#> 16  016  2.148957351  0.16239276
#> 17  017  0.677080730 -0.60584486
#> 18  018  0.697203691 -0.74054539
#> 19  019 -0.629835438 -0.12381716
#> 20  020 -0.239803183  0.93215567
#> 21  021  2.388829859  1.06816430
#> 22  022 -0.631382235 -0.76027081
#> 23  023  1.074293999  1.43566987
#> 24  024 -0.096146373  0.23967926
#> 25  025  2.326556263  0.47613575
#> 26  026  1.225405118 -0.11018272
#> 27  027  1.039936279  1.02934306
#> 28  028  0.294940225  0.14815138
#> 29  029  1.274258820  0.82135794
#> 30  030  1.405004392  0.99019289
#> 31  031  0.046810432 -0.89108019
#> 32  032  1.301615939  1.04455273
#> 33  033 -1.890321167 -1.17058080
#> 34  034 -0.573464180 -0.80814654
#> 35  035 -1.357927654 -0.72393060
#> 36  036  0.600261506 -0.10994652
#> 37  037  0.375300584  0.50883235
#> 38  038 -0.445425019 -0.47855558
#> 39  039  1.337330649  1.19170873
#> 40  040 -0.429563655  0.47756676
#> 41  041  0.500625247  0.85458508
#> 42  042 -0.093452079 -0.44784638
#> 43  043 -0.357828763 -0.85207813
#> 44  044 -1.023692181 -0.11665870
#> 45  045 -1.483624845 -0.41454011
#> 46  046 -0.559753959  0.10492399
#> 47  047  1.842715782  0.44563202
#> 48  048  0.187787571  0.93231060
#> 49  049  1.455045240  0.17639744
#> 50  050  1.162066139  0.55047489
#> 51  051  0.975359248  0.64231442
#> 52  052  0.604312887  1.02489856
#> 53  053  1.500965681 -0.14433864
#> 54  054  0.301277941 -0.23803807
#> 55  055  0.070814160 -1.43444989
#> 56  056  0.244707974 -0.56156406
#> 57  057  0.072939961  1.00335730
#> 58  058  1.766884273  0.16482765
#> 59  059 -0.824107655 -0.75055288
#> 60  060  2.153782017  0.55343007
#> 61  061  1.492297639 -0.06983878
#> 62  062 -0.334394545 -0.91437384
#> 63  063  0.767630616  0.56426720
#> 64  064  0.559412616  0.57424039
#> 65  065 -1.158287669 -0.68909736
#> 66  066 -0.202056026  0.26818500
#> 67  067  0.007068973  0.82758686
#> 68  068 -2.471714818 -0.45546364
#> 69  069 -0.828233044  1.13857556
#> 70  070 -0.249874979 -0.93490741
#> 71  071  0.624052951  1.59238265
#> 72  072  0.151079977  1.21148591
#> 73  073  0.302194377  1.46807744
#> 74  074  0.843644273  0.99579810
#> 75  075 -1.603040328 -0.92020699
#> 76  076 -0.898101938 -0.40567324
#> 77  077 -0.863081279 -0.21218620
#> 78  078 -0.844007776 -1.11944031
#> 79  079  2.419845091  1.84730678
#> 80  080 -0.381398287 -0.01042072
#> 81  081  0.058393629 -0.50407058
#> 82  082 -1.024396884 -0.84574680
#> 83  083  0.093315639 -0.99873580
#> 84  084  0.411380440  0.58021780
#> 85  085  1.419692298  0.92322236
#> 86  086  1.436040245  0.12399008
#> 87  087 -1.817086238 -1.67975122
#> 88  088  0.031852962 -0.94348623
#> 89  089  0.788733882 -0.81406615
#> 90  090 -0.706065525 -1.18613692
#> 91  091 -0.813393365 -0.99215827
#> 92  092  0.694125229  0.11371906
#> 93  093  0.393249385  0.13789163
#> 94  094 -0.914252069  0.26636588
#> 95  095  0.372276802  0.94849920
#> 96  096  2.518648773  2.09678077
#> 97  097 -0.989319017 -0.07385438
#> 98  098  0.263279417 -0.13386036
#> 99  099 -0.271282378  1.38765939
#> 100 100  0.605262909 -1.03304846

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.217919636 -1.10646423
#> 2   002  0.806691538 -0.88128045
#> 3   003  0.986632630  0.49079891
#> 4   004 -1.436589595 -0.63632568
#> 5   005 -1.239370633 -1.46414469
#> 6   006 -0.372548613 -1.09533859
#> 7   007 -0.465901665 -0.02317672
#> 8   008 -0.878715177 -0.33356916
#> 9   009  1.926361195  1.66577213
#> 10  010  0.459110123  0.46652128
#> 11  011 -0.736466310 -1.78862694
#> 12  012  2.048479158  1.46491285
#> 13  013 -0.566653918 -2.50799841
#> 14  014  1.106712396  2.91474862
#> 15  015  0.867494210  1.14519315
#> 16  016 -0.160108923 -1.10479684
#> 17  017  0.978047734  0.81565177
#> 18  018  2.027790025  1.39163582
#> 19  019  0.206088474 -0.52202769
#> 20  020 -1.288663743 -0.74241633
#> 21  021 -1.921390198 -0.92265655
#> 22  022 -0.907387947  1.66320401
#> 23  023 -0.388621232 -0.49739484
#> 24  024  1.659065836  0.20518196
#> 25  025 -1.008035348  0.58774160
#> 26  026 -0.799896364 -0.39964182
#> 27  027 -0.651143714 -0.47861365
#> 28  028  0.785377067  1.29186475
#> 29  029 -0.401214657  0.51356589
#> 30  030  1.167998866 -0.67910089
#> 31  031 -0.237487141 -1.35597096
#> 32  032  0.496308108  0.93069655
#> 33  033 -0.369841091 -1.05621672
#> 34  034  0.315594262  0.25211854
#> 35  035 -0.817913654 -1.15242175
#> 36  036  1.991274253  1.24186028
#> 37  037 -0.591745345 -1.18089337
#> 38  038  1.179684783  0.47884438
#> 39  039 -0.791439461 -1.40383532
#> 40  040  0.367775064 -0.24282126
#> 41  041  2.124637042  1.10808123
#> 42  042 -0.996938511 -0.32087338
#> 43  043 -0.864477487 -0.71939575
#> 44  044 -0.046586558 -1.31039444
#> 45  045  1.040863787 -0.50508656
#> 46  046 -0.286074804  0.83971764
#> 47  047 -1.025417362 -0.07517360
#> 48  048  0.011525212  0.80427299
#> 49  049  1.330455676 -0.42839267
#> 50  050  0.007409148  1.15351765
#> 51  051  0.873501105  2.00458441
#> 52  052  2.588914721  1.31631032
#> 53  053  1.061756048 -0.09046823
#> 54  054  3.049241025  0.53433227
#> 55  055 -1.507501702 -1.86406543
#> 56  056 -0.749100515 -2.13040192
#> 57  057 -3.230385530 -1.12955792
#> 58  058  0.430170373  0.86152884
#> 59  059  0.571748441  1.41355038
#> 60  060  0.944813208  0.61086325
#> 61  061 -0.506739749 -0.52997020
#> 62  062  1.962342482  1.57200308
#> 63  063  0.215399779  0.57266073
#> 64  064  0.805724302  0.80533459
#> 65  065 -0.262500695 -0.06505177
#> 66  066  0.899916632  1.74734105
#> 67  067 -0.781316958  0.42896497
#> 68  068 -0.661025455  0.15123930
#> 69  069  0.464455813 -0.64696783
#> 70  070  0.808142894 -0.55072368
#> 71  071 -1.769404914  0.50447344
#> 72  072  1.222054057  0.52141970
#> 73  073  1.389517725 -0.32930858
#> 74  074  0.848846809  0.33158285
#> 75  075  0.537277338  0.41941063
#> 76  076  1.207000948  0.44115897
#> 77  077 -0.287377344 -1.07328735
#> 78  078  1.418620827  1.88959004
#> 79  079  0.408421223  0.69195298
#> 80  080  1.997684461  1.49309849
#> 81  081  1.047791095  1.09774259
#> 82  082 -0.680975602  0.02955394
#> 83  083  1.262568678  0.03069817
#> 84  084  1.226470980  0.54280594
#> 85  085  1.064552356  0.92961749
#> 86  086 -2.000991667 -2.04103905
#> 87  087  1.656158070 -0.72501446
#> 88  088  0.195562426  0.70755889
#> 89  089  0.434395741  1.28350969
#> 90  090 -0.760928909  0.68172105
#> 91  091 -1.657210530 -0.11874289
#> 92  092  1.711881012  0.59340192
#> 93  093  0.835822946  0.76573891
#> 94  094  0.198991869  0.03932598
#> 95  095  0.034461876  1.17058083
#> 96  096 -0.262957625 -0.18425388
#> 97  097  0.607351218 -1.49089833
#> 98  098  0.268561099 -0.19997515
#> 99  099  1.005739691 -0.72475675
#> 100 100 -0.435599046  0.56610827