Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4388085
#> Y_2 0.4388085 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -1.19987831  1.1715355
#> 2  0.49469888  0.2064299
#> 3 -1.10100347  0.4928411
#> 4  0.09886246  1.1105524
#> 5 -1.70534841 -0.8490287

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.97282687  1.46477053
#> 2   002 -1.69697650 -2.10813317
#> 3   003  0.98335417  0.09846184
#> 4   004 -0.84519082 -1.43215217
#> 5   005 -0.62123321  0.58510509
#> 6   006 -0.18232695 -0.56761788
#> 7   007 -0.50729884 -1.41557938
#> 8   008  2.13744539  2.56983584
#> 9   009  0.26359608 -1.40772191
#> 10  010 -1.71850379 -0.16634213
#> 11  011 -2.66715998 -1.77351327
#> 12  012  0.40375219  0.61258143
#> 13  013 -0.55107862 -0.33670514
#> 14  014  0.08021543  0.08222536
#> 15  015  1.12949925  0.25037153
#> 16  016 -0.36356375 -0.09296453
#> 17  017  0.22373480 -0.64728548
#> 18  018 -1.41837895  0.41973961
#> 19  019 -1.19565055 -0.34984438
#> 20  020 -0.04741737 -1.01449773
#> 21  021 -0.26159137  0.41650835
#> 22  022  1.69474592 -0.17897204
#> 23  023  1.97485815  0.39696743
#> 24  024 -0.96916855 -1.00914374
#> 25  025  0.14312812 -1.22980750
#> 26  026 -0.15628119 -1.37668926
#> 27  027  0.59267100 -0.70862062
#> 28  028  0.21014238 -0.13441788
#> 29  029  0.75602315  0.63305660
#> 30  030  0.26213599  0.24510079
#> 31  031 -0.06695287  0.30268445
#> 32  032 -0.35018382 -0.17949401
#> 33  033 -1.61249303 -2.41470578
#> 34  034 -0.13618307 -0.11775115
#> 35  035 -1.26502156 -1.12562486
#> 36  036 -1.49036847 -0.89602280
#> 37  037 -0.54460482 -0.55163349
#> 38  038  0.27150627 -0.22984151
#> 39  039 -2.76583874 -1.64139667
#> 40  040  0.38196044 -1.11972141
#> 41  041 -1.81033164 -0.26670763
#> 42  042  0.20331101  0.43013517
#> 43  043  0.79561104  0.62461443
#> 44  044  0.02789253 -0.80211917
#> 45  045  0.75533501 -0.17471458
#> 46  046 -0.18202260 -1.12583522
#> 47  047  0.41654158 -1.51391017
#> 48  048  0.99545082  0.58132297
#> 49  049  0.55198929  0.20805053
#> 50  050 -0.75294958  0.21185553
#> 51  051 -0.80832989 -0.63243021
#> 52  052  0.78318627  0.55346897
#> 53  053  0.50369192  1.43058853
#> 54  054  1.16976626 -0.56131749
#> 55  055  0.85856712  0.33641978
#> 56  056 -1.29380247 -1.31324350
#> 57  057  1.36007881  1.84421622
#> 58  058 -0.45732943  0.54215810
#> 59  059  0.42714681 -0.46269404
#> 60  060  0.98728555  0.44376730
#> 61  061 -0.59137012 -0.42411318
#> 62  062  1.23434773  1.27687077
#> 63  063  0.51251914  1.06758647
#> 64  064  1.79565256  0.20764881
#> 65  065  0.03430961  0.60857388
#> 66  066 -1.01470900 -1.67183692
#> 67  067 -0.94047267 -1.38507480
#> 68  068  0.04642361  0.62488155
#> 69  069 -1.03730336 -0.66753794
#> 70  070 -0.50325943 -0.93763645
#> 71  071  0.99933000  1.45285147
#> 72  072  0.20851280  1.17520496
#> 73  073 -0.31458130 -1.14675506
#> 74  074 -1.14421362  0.82317068
#> 75  075 -0.97448918  0.97682875
#> 76  076 -0.38552415  0.20456908
#> 77  077  1.28977817  1.15552047
#> 78  078  0.61782940 -0.04518648
#> 79  079  0.11386575  1.04108483
#> 80  080  1.13422642 -0.08795449
#> 81  081  0.62599173 -0.26522308
#> 82  082 -0.74303480 -0.37949297
#> 83  083 -1.92540757  0.34271949
#> 84  084 -1.85861563  0.84559005
#> 85  085 -0.41826759 -1.55820976
#> 86  086  0.92825756  0.63297812
#> 87  087 -1.68792305 -0.10162712
#> 88  088 -1.14409779  0.66092027
#> 89  089  0.07059731  1.47097120
#> 90  090  0.36127635 -0.16707852
#> 91  091 -1.15757501  0.09629951
#> 92  092 -0.01762855 -1.63814350
#> 93  093 -1.40645494 -1.78607805
#> 94  094 -1.91934403 -2.21050615
#> 95  095 -0.11642125  1.06657605
#> 96  096  0.10410214 -0.43829138
#> 97  097 -0.48327003 -0.83224441
#> 98  098 -1.22212924  1.11981267
#> 99  099  1.75806862  1.65202484
#> 100 100 -0.66266215 -0.91939021

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.35566001 -0.088163321
#> 2   002 -1.33922392 -0.025564182
#> 3   003  0.31878574 -1.334628738
#> 4   004 -0.77983522 -0.309275871
#> 5   005 -0.48533715 -0.416793730
#> 6   006  0.43817930  1.007587371
#> 7   007 -2.06311815 -1.125706480
#> 8   008  0.66723294  0.480444699
#> 9   009 -0.71436866 -0.520035485
#> 10  010 -1.90889786 -0.010698420
#> 11  011  0.14581421 -1.331955361
#> 12  012  1.33945843  1.308890418
#> 13  013  0.86611425  1.376495254
#> 14  014  0.08148771 -0.915597262
#> 15  015  0.09290127  0.762866568
#> 16  016  0.27947547 -0.491592685
#> 17  017  1.21623294  0.558523265
#> 18  018 -0.97522096 -1.264790738
#> 19  019  2.57295267  1.323008263
#> 20  020 -1.98138461 -0.902850389
#> 21  021 -1.77096655 -0.871147675
#> 22  022  0.37151324  0.706636608
#> 23  023  0.22601842  0.591091334
#> 24  024  0.33895393  0.138558125
#> 25  025 -1.04669295 -0.551347448
#> 26  026 -1.81691502  0.780664051
#> 27  027  1.92717252 -0.577735059
#> 28  028 -0.11057112 -1.277270472
#> 29  029  1.68022284 -0.723121898
#> 30  030  0.03828424  0.523239949
#> 31  031  0.06240301  0.385156059
#> 32  032  0.05527382 -1.313728499
#> 33  033 -1.45879137 -1.071410586
#> 34  034  0.28717492 -0.510849876
#> 35  035 -1.77296258 -1.992249720
#> 36  036  0.74650767  1.821706469
#> 37  037 -0.98676989 -1.183524938
#> 38  038 -0.94691371 -0.227492559
#> 39  039  1.02614719  0.935932157
#> 40  040 -1.14786471 -0.261708293
#> 41  041  0.86895515  0.106404862
#> 42  042 -0.02222362  0.602412383
#> 43  043  0.42900418  0.126905365
#> 44  044  0.20736303 -1.056925112
#> 45  045 -0.11683287 -0.308772731
#> 46  046 -0.05802879 -0.828142923
#> 47  047 -0.25861305  2.100401213
#> 48  048  1.92274038  0.647374814
#> 49  049  0.70698473 -0.628697041
#> 50  050  0.61960291  1.944794544
#> 51  051  0.78921806  1.050600312
#> 52  052  0.59579708  0.661858312
#> 53  053  2.17656799  0.772080881
#> 54  054 -1.13457311 -1.139739087
#> 55  055  0.75148285 -1.398228389
#> 56  056  0.88644098  1.441963109
#> 57  057  0.03456842 -1.104417897
#> 58  058  0.46281488 -0.788568376
#> 59  059  0.23946070  0.780001662
#> 60  060 -0.28230654 -0.417411499
#> 61  061  1.41623166  1.567090715
#> 62  062 -0.52027340 -1.298297491
#> 63  063 -1.24855759 -0.378966797
#> 64  064  0.25516482 -0.115422952
#> 65  065 -1.83963291 -0.703126613
#> 66  066  0.60692405  0.535849106
#> 67  067 -0.30064755  0.546717841
#> 68  068 -1.33451285  0.195919956
#> 69  069  0.21977304  0.567254376
#> 70  070 -0.91530890 -1.447171185
#> 71  071  0.88071274  0.289583004
#> 72  072  1.10689125  0.855091612
#> 73  073  1.90311103  0.951372512
#> 74  074 -0.52712746 -0.298339701
#> 75  075  0.82130391  0.421158833
#> 76  076  1.48214103 -0.256578000
#> 77  077 -0.63355035 -0.233355093
#> 78  078 -1.71522681  0.656356952
#> 79  079  1.46265616  2.360428084
#> 80  080 -1.07468872  0.257866662
#> 81  081 -1.45913850 -0.031099094
#> 82  082 -1.15349193 -0.005134086
#> 83  083  0.28680987 -0.632553724
#> 84  084 -0.58430612 -1.672159099
#> 85  085 -0.27110976 -0.995958398
#> 86  086  0.15941651 -0.795569600
#> 87  087 -0.73640969  1.727883907
#> 88  088 -0.56054145 -1.118795058
#> 89  089  0.24323042  1.302298370
#> 90  090  0.90861645  1.536274663
#> 91  091  0.25898883 -0.465096455
#> 92  092 -1.78562210 -0.644530741
#> 93  093 -0.13903913 -0.145215898
#> 94  094  1.23403311  1.192873902
#> 95  095 -0.30824263  0.757772984
#> 96  096 -0.29690323 -1.438237340
#> 97  097 -0.80081787 -0.989929497
#> 98  098 -1.47269222 -0.121359923
#> 99  099 -2.15606640 -0.464363127
#> 100 100 -1.38988365 -0.218722534