Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4102785
#> Y_2 0.4102785 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.7865441  0.1914098
#> 2  0.1198059 -0.2424647
#> 3 -2.4782708 -0.8689023
#> 4  1.5157666  0.1165733
#> 5  1.7948809  1.7749713

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.048713040 -0.786177632
#> 2   002  0.888537788  0.308998067
#> 3   003 -0.137355045  1.493419469
#> 4   004  0.806910699  1.821630233
#> 5   005  0.285035289 -0.052660218
#> 6   006  0.810716804 -0.637200462
#> 7   007 -0.564846436  1.053644006
#> 8   008 -0.180337866  0.234326489
#> 9   009 -0.464297931 -1.230116820
#> 10  010  2.256000970  1.130966855
#> 11  011 -0.439702622 -0.601858965
#> 12  012 -1.128687061 -0.480409173
#> 13  013 -0.088274998 -0.037358988
#> 14  014  0.754624642  2.343797143
#> 15  015 -1.247106251 -0.417486522
#> 16  016 -1.458363196 -0.113933675
#> 17  017 -0.443097450  0.925845136
#> 18  018  0.899032702 -0.076932036
#> 19  019 -1.055549780 -0.112793845
#> 20  020 -0.176677393 -0.598160711
#> 21  021 -0.382471240 -1.586632288
#> 22  022 -0.164224580  0.486223792
#> 23  023 -1.693542324  0.888309194
#> 24  024  0.650297159  0.001868557
#> 25  025 -1.945287451 -1.394004871
#> 26  026  0.594069693  0.603877211
#> 27  027 -0.494925011 -0.580257698
#> 28  028  2.211601075  1.469622202
#> 29  029 -0.674043706 -0.508992463
#> 30  030 -1.470365723 -0.690322965
#> 31  031 -1.036108422 -1.168790064
#> 32  032  2.553850237  0.656830235
#> 33  033 -0.206691580  0.712339815
#> 34  034  0.102617448  1.037400591
#> 35  035 -0.244997636 -0.088525448
#> 36  036 -1.019164449  0.420609338
#> 37  037  0.463221970  0.215591027
#> 38  038  0.535289134  0.141326072
#> 39  039 -0.256506158 -0.593019759
#> 40  040 -3.531285180 -1.781643495
#> 41  041  2.032060537  3.626975007
#> 42  042 -0.454872370  0.202841449
#> 43  043  0.327108663 -0.286586493
#> 44  044 -0.326478212 -1.291897338
#> 45  045 -0.028326881 -0.453869379
#> 46  046  0.557188186  0.908652880
#> 47  047  1.169328389 -0.406048635
#> 48  048  0.100891333 -0.868647768
#> 49  049 -2.262584744 -3.347513053
#> 50  050 -1.217650017 -0.246049706
#> 51  051 -0.712579146 -0.617808255
#> 52  052  0.337706331  0.595597081
#> 53  053 -2.076399910 -0.095630925
#> 54  054 -1.412584890 -1.087894011
#> 55  055  0.002814547  0.039404832
#> 56  056  0.399648964  0.988580446
#> 57  057 -0.471664746 -0.279999541
#> 58  058 -0.285945005 -0.111912528
#> 59  059 -0.345986734  0.227539866
#> 60  060  0.206703645  1.664596837
#> 61  061  0.453360439  1.463013525
#> 62  062 -1.428580520 -1.151796883
#> 63  063  0.154635372 -0.656314277
#> 64  064 -1.890427944 -1.211771432
#> 65  065 -0.760899947  1.401114518
#> 66  066  0.258350762 -0.606274716
#> 67  067 -0.529420439  0.826120980
#> 68  068 -1.313073156 -0.196906222
#> 69  069 -0.703067795  0.189512749
#> 70  070  0.861258707  1.031124683
#> 71  071  0.740956693 -0.317816792
#> 72  072  0.080132917 -0.523419097
#> 73  073 -0.853001992 -0.714863663
#> 74  074  1.218650526  1.262879882
#> 75  075  0.164711375 -0.335460656
#> 76  076  0.388866866 -0.553169955
#> 77  077 -0.521093104 -0.311147593
#> 78  078 -1.720069745 -1.636017630
#> 79  079  0.235495964 -0.521516592
#> 80  080  0.631340831 -0.401828070
#> 81  081  0.370140248 -0.718795870
#> 82  082 -0.652577836 -0.628053232
#> 83  083 -1.019676274  0.794849811
#> 84  084 -0.585767844 -0.647714356
#> 85  085 -0.559173809 -0.033071011
#> 86  086 -0.413744861 -1.602009180
#> 87  087  0.692714030  0.774219270
#> 88  088 -0.094830699 -0.622226007
#> 89  089 -0.041317747  0.062186677
#> 90  090 -0.184514348  0.200075351
#> 91  091  0.163913856 -0.934159811
#> 92  092  0.370079467 -0.011919820
#> 93  093 -2.154239687 -1.649643906
#> 94  094 -0.655163111 -0.204198261
#> 95  095 -1.840477875 -2.479042795
#> 96  096  0.390507610  0.342699816
#> 97  097  0.168539577 -1.191781610
#> 98  098 -0.956658735  0.059498499
#> 99  099 -0.373977275 -0.289766014
#> 100 100 -1.470003266 -1.860257229

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.120305447  0.26191364
#> 2   002  1.684685904  0.37644578
#> 3   003 -0.157392583  0.13017867
#> 4   004 -0.323039034  0.12594096
#> 5   005 -0.044423640 -0.91125305
#> 6   006 -1.020196682 -0.63531818
#> 7   007  0.224342167 -0.57863093
#> 8   008 -1.149200140  1.08817907
#> 9   009  1.911818668  0.78949966
#> 10  010 -1.725256411 -1.06957620
#> 11  011 -0.018081885  1.22215103
#> 12  012  0.884906787  0.07905823
#> 13  013 -0.663091187 -0.51559932
#> 14  014 -0.167534632  0.50927222
#> 15  015  1.709114932  0.95957403
#> 16  016  0.615983152  0.73549215
#> 17  017  0.568865496 -0.90573997
#> 18  018  0.365994095 -0.65033477
#> 19  019  0.130315045  0.64096476
#> 20  020  0.458989788 -0.11556044
#> 21  021  0.758757613  0.52078386
#> 22  022  2.090795491  1.28570635
#> 23  023 -0.348965823 -0.90382766
#> 24  024  0.766294001  0.12508430
#> 25  025  0.511981817  0.19127916
#> 26  026  0.418900782 -0.84819566
#> 27  027  0.335013020  0.45454462
#> 28  028 -0.522026862 -0.70541700
#> 29  029  0.750863023 -0.38944711
#> 30  030  0.673534168  0.41985714
#> 31  031  1.734810872 -0.26015092
#> 32  032 -0.271971586  0.99999394
#> 33  033 -1.975594908 -0.35074039
#> 34  034 -1.346698633 -0.29077461
#> 35  035  1.240080751  1.14277851
#> 36  036  0.916164132  0.14900833
#> 37  037 -1.293451245 -0.83247783
#> 38  038  0.379169795 -0.33515533
#> 39  039  2.150707834  1.81687025
#> 40  040  0.931958002  0.42927202
#> 41  041  0.949656662  0.36497618
#> 42  042  0.020181401 -0.61421766
#> 43  043  0.393618886 -0.30652853
#> 44  044 -1.886659606 -0.33680817
#> 45  045  0.543710386  0.81353397
#> 46  046 -1.321363827 -2.56962894
#> 47  047 -0.365891413  0.47791195
#> 48  048  0.187997628  1.54096706
#> 49  049  0.171215513 -1.36147180
#> 50  050 -1.300879777 -1.52839624
#> 51  051  0.495621801  0.40986104
#> 52  052  1.151896532  0.67245016
#> 53  053 -0.648377335 -1.28646583
#> 54  054 -1.051186869  0.37410283
#> 55  055 -0.886050531 -0.32629319
#> 56  056  0.229579197  0.18816426
#> 57  057  1.247636482  1.51936513
#> 58  058  1.727662193  0.05822599
#> 59  059 -0.631307204 -0.56122598
#> 60  060  0.614190457 -1.40771411
#> 61  061 -0.414453259 -1.17068320
#> 62  062  0.152777926  0.63803989
#> 63  063  0.005372561  1.39787245
#> 64  064 -0.851790521 -1.33441498
#> 65  065 -1.197891070  0.28920941
#> 66  066  0.756579849  0.48754387
#> 67  067  0.761445162  1.50597715
#> 68  068  2.973942368  1.71598800
#> 69  069 -0.400811451 -0.20084687
#> 70  070 -0.846745715 -0.43469084
#> 71  071  0.837860966  0.11713333
#> 72  072  1.315332310  0.80462967
#> 73  073 -0.375572024 -0.78920539
#> 74  074  1.766408451  0.60981461
#> 75  075  0.193073595 -1.13902263
#> 76  076  0.774634536  1.81118194
#> 77  077  0.335508652 -0.26598240
#> 78  078 -1.145500521 -0.84020278
#> 79  079  0.456608555  0.02852407
#> 80  080  1.419996139 -0.03759701
#> 81  081  1.060478136 -0.44690856
#> 82  082 -1.912033029  0.22239144
#> 83  083 -1.432715857 -0.50503936
#> 84  084 -0.779583547 -0.50736149
#> 85  085  0.025005980 -0.98455713
#> 86  086 -1.471861592  0.78518015
#> 87  087  0.088536817 -0.29626946
#> 88  088 -0.248072021  0.03763931
#> 89  089 -0.362921139  0.80678939
#> 90  090  0.600489022 -0.19330062
#> 91  091 -0.820741797  0.29253274
#> 92  092  0.284524510 -0.14998969
#> 93  093 -1.703969601  1.05368155
#> 94  094  2.733713435  0.59023046
#> 95  095 -0.821832379  0.25840731
#> 96  096  0.263453633  0.66392395
#> 97  097  1.861788898  0.23980312
#> 98  098  1.114270475  0.44183083
#> 99  099  0.028127524  0.55743171
#> 100 100  2.206730713  1.89629428