Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5433141
#> Y_2 0.5433141 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.4876730  0.6496364
#> 2  0.3213183 -0.2918194
#> 3  0.4081731  0.6914504
#> 4 -1.6783266 -1.9124188
#> 5  0.8165866 -0.5600976

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.481076930 -0.46647115
#> 2   002 -0.262098619 -0.82898731
#> 3   003  0.614383040  0.81197941
#> 4   004 -1.104529717 -1.36429648
#> 5   005  1.303050731  0.08734576
#> 6   006  0.195018011  0.28458123
#> 7   007 -0.412817900  0.31115272
#> 8   008  0.470388331 -0.23699354
#> 9   009  0.199359828 -0.78646229
#> 10  010  0.125863809 -1.11033845
#> 11  011  0.908429692  0.72990871
#> 12  012  0.707066510  0.14423814
#> 13  013 -0.269050378  0.70368175
#> 14  014  0.657253821 -1.28909020
#> 15  015  0.411091273  0.44148958
#> 16  016  1.348828533  1.22314841
#> 17  017  0.648945278  1.27883753
#> 18  018  0.904260373  2.00273428
#> 19  019 -2.100384563 -1.38412590
#> 20  020  0.811412140  1.46138965
#> 21  021  1.327712499  2.80137303
#> 22  022 -2.085713118 -1.78574151
#> 23  023  0.536335565  0.38926182
#> 24  024  2.268940868  1.55616326
#> 25  025 -0.962504066 -0.41303580
#> 26  026 -1.175882008 -2.23825507
#> 27  027  0.189188737  0.50610298
#> 28  028  0.291589274  0.19489684
#> 29  029  1.752562542  0.21013148
#> 30  030  0.537249754  0.37181746
#> 31  031 -0.154531742 -0.41677653
#> 32  032 -0.122119890  0.39526426
#> 33  033 -0.991496491  0.74716436
#> 34  034  0.338130984 -0.15319367
#> 35  035 -1.927280580  0.29848235
#> 36  036 -0.744160474  0.02384081
#> 37  037 -0.313919554  0.08829199
#> 38  038 -0.023408220 -1.10530673
#> 39  039 -0.416098425 -0.74677368
#> 40  040  0.710508553  0.52443462
#> 41  041 -0.405186130  0.27010570
#> 42  042 -0.531489423 -1.01490196
#> 43  043  1.497776219  0.09613354
#> 44  044  0.430308310  0.46883097
#> 45  045 -1.767064481 -1.45429707
#> 46  046 -1.562499657 -1.17259412
#> 47  047 -0.552271831  0.18289738
#> 48  048 -0.524305227  0.65416354
#> 49  049  0.193067099 -0.63286865
#> 50  050  1.032796719  0.60394344
#> 51  051  0.232915360 -0.75367530
#> 52  052 -1.427849856 -0.34072097
#> 53  053  0.066285028 -0.47527089
#> 54  054  1.696688309  1.23625608
#> 55  055 -0.013372316 -0.23495033
#> 56  056 -1.614358573  0.10751067
#> 57  057 -1.207335551 -0.27961456
#> 58  058 -0.076193395 -0.36496384
#> 59  059 -1.458306049 -0.82173956
#> 60  060  2.034246314  2.34851909
#> 61  061  1.431751914 -0.48955418
#> 62  062  0.915319374 -0.39214958
#> 63  063  0.046356567 -0.73462976
#> 64  064 -0.375819749  1.35956551
#> 65  065  1.447820049  0.74842491
#> 66  066  1.700069320  0.94944772
#> 67  067  0.654380620 -0.92746589
#> 68  068  0.167560417  0.63672977
#> 69  069 -1.170708493  0.25624980
#> 70  070 -0.923449493 -0.67439205
#> 71  071 -1.116232539  0.09184994
#> 72  072  0.733422258 -1.36896415
#> 73  073  0.129175619 -0.15105223
#> 74  074 -0.480588127 -0.26815915
#> 75  075  0.958981409  0.05274869
#> 76  076 -0.110124454  0.16214227
#> 77  077  0.478716225 -0.24340034
#> 78  078  0.504641560  0.44683366
#> 79  079 -0.181703772 -0.61771886
#> 80  080 -0.002003394  1.24990848
#> 81  081  1.089449635  1.56733641
#> 82  082  0.488397865  1.18855890
#> 83  083  0.391722690 -0.56321889
#> 84  084 -0.098445718 -0.03370982
#> 85  085 -0.980656555  0.51018606
#> 86  086 -0.319998807  0.16275190
#> 87  087  0.167525900 -0.46414368
#> 88  088 -0.853878824 -0.64625798
#> 89  089 -0.362133670  0.07470424
#> 90  090 -0.024526381 -0.90635666
#> 91  091 -0.828743059 -2.71829635
#> 92  092  0.795552834 -0.21458803
#> 93  093  0.770821163  0.71427408
#> 94  094 -0.883032767  1.11036593
#> 95  095 -0.442269162 -0.00807391
#> 96  096 -0.347950045  1.09459736
#> 97  097 -0.306586714 -0.19801237
#> 98  098  0.065324362  0.71981959
#> 99  099  0.778708779  0.13625507
#> 100 100  0.741031789 -0.63436474

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.35421600  1.219593863
#> 2   002  0.32916547  1.089833759
#> 3   003 -0.69073384 -0.664548853
#> 4   004 -1.74508926 -1.360229616
#> 5   005  1.02529441 -1.157606495
#> 6   006 -0.89546402 -1.024330829
#> 7   007  0.08961897  0.876969910
#> 8   008  1.01105022  1.115579708
#> 9   009  1.07985552  1.038073818
#> 10  010 -0.31501349 -0.117341748
#> 11  011 -1.30084037 -0.013993167
#> 12  012  0.79864027  0.089919166
#> 13  013  1.67271116  0.147193537
#> 14  014  1.93465083  0.317239946
#> 15  015  1.00003870  0.069341757
#> 16  016 -0.56069472 -1.212078831
#> 17  017  1.34470126 -0.029435308
#> 18  018 -1.15541533 -0.480354891
#> 19  019  0.32151683  1.442026385
#> 20  020 -1.04795115 -1.890619220
#> 21  021 -0.92132555 -0.121017229
#> 22  022 -0.31012001 -0.209035068
#> 23  023 -0.08084939 -2.030609528
#> 24  024 -0.71992159  0.163584958
#> 25  025 -0.50839437  0.087695046
#> 26  026 -1.40285415 -1.224229586
#> 27  027 -0.17311873 -0.644899367
#> 28  028  1.07654500  0.609254942
#> 29  029  0.45649960  0.118277229
#> 30  030  0.80889568  1.511252072
#> 31  031 -0.30104119 -1.287046913
#> 32  032 -0.04196507 -0.240036017
#> 33  033 -0.31920888 -1.166376117
#> 34  034  0.91373907  1.281035144
#> 35  035  0.67777815  1.240912001
#> 36  036  0.17393245  0.706512397
#> 37  037 -1.68537336 -0.677125855
#> 38  038  1.24266410  0.662333508
#> 39  039 -0.45922943 -0.657079872
#> 40  040 -1.72447259  0.451689900
#> 41  041 -0.60895822 -1.400467601
#> 42  042 -2.57829345 -2.008196101
#> 43  043  1.36538694 -0.311236257
#> 44  044  1.35138710  1.509920487
#> 45  045  0.97863311 -0.551740247
#> 46  046  0.64145908  0.429550451
#> 47  047 -0.19320457 -0.578796208
#> 48  048  0.56627439 -0.087137266
#> 49  049 -1.01362798 -0.784307248
#> 50  050  0.62339169  1.136632957
#> 51  051  1.65209484 -0.371446380
#> 52  052 -0.77248926 -0.919276733
#> 53  053 -1.87383004 -0.388337314
#> 54  054 -1.39306900 -0.929790048
#> 55  055 -1.27094203 -1.741051943
#> 56  056  1.01834217 -0.447029784
#> 57  057 -0.64253760 -0.192223987
#> 58  058  0.38983653 -0.006074293
#> 59  059  0.23316596 -0.380606704
#> 60  060 -0.23621285  0.142216333
#> 61  061  1.36502771  2.442964135
#> 62  062  0.16377556  0.014175086
#> 63  063 -0.26436490  0.152124800
#> 64  064 -0.74920532 -1.989689262
#> 65  065  1.42121123  0.730298395
#> 66  066 -0.85165053 -1.141662564
#> 67  067  0.69418508 -1.687232260
#> 68  068  0.72782686  0.983848822
#> 69  069 -1.16346429  0.179358855
#> 70  070 -0.86248563 -0.910863786
#> 71  071  0.98852709  2.034225288
#> 72  072  0.74654085  0.925425676
#> 73  073  2.43004427  1.660309351
#> 74  074  0.21935385 -0.621020040
#> 75  075 -2.60318244 -1.029715871
#> 76  076 -0.19733563  0.666643262
#> 77  077  0.46099750  0.972287273
#> 78  078  0.25369578 -0.016691178
#> 79  079 -0.28369730 -0.353386377
#> 80  080  0.12834126 -0.392802854
#> 81  081  0.74526292  0.457914638
#> 82  082  1.05339377 -0.363074674
#> 83  083  0.99845582  0.618885951
#> 84  084 -1.48484884  0.450246564
#> 85  085  1.61721832  0.711518541
#> 86  086  0.80676344  0.959341503
#> 87  087 -0.78285967 -1.066976583
#> 88  088  0.33455067 -0.256050544
#> 89  089  1.01894617  0.304557464
#> 90  090 -0.32134063  0.596713991
#> 91  091 -0.52006257 -1.090851123
#> 92  092 -1.31738513 -0.199258411
#> 93  093  0.65566568  0.780433478
#> 94  094  1.79643850 -0.389466020
#> 95  095 -1.74976758 -1.227437003
#> 96  096  0.52476054 -0.142537426
#> 97  097 -1.58055429 -1.846100327
#> 98  098  0.87186171  2.275695123
#> 99  099  1.94149182 -0.069723777
#> 100 100  0.71705841  2.172177622