Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4904787
#> Y_2 0.4904787 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.6310861  2.5437741
#> 2 -0.7592917 -1.3060524
#> 3 -0.2230369 -1.0305784
#> 4 -1.0101462  0.7766278
#> 5 -0.1249788 -0.3665335

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.052422835  0.301197014
#> 2   002  0.286773575  0.463584749
#> 3   003  0.937932705  0.497902889
#> 4   004 -0.796300055 -0.600292188
#> 5   005 -0.548073495 -0.284272358
#> 6   006 -0.310628669 -0.610057442
#> 7   007  0.383538693  0.197750442
#> 8   008  0.954615958 -0.076984991
#> 9   009  0.354897403  1.252892940
#> 10  010  0.546446567 -1.554955796
#> 11  011 -0.063015330 -0.235214646
#> 12  012  0.305583305 -1.132534487
#> 13  013  0.343977193  0.599012701
#> 14  014 -0.477150760  0.521502630
#> 15  015  0.888145017  0.703552529
#> 16  016 -0.773010155  0.059700945
#> 17  017  3.160753692  0.578510818
#> 18  018  0.247601407 -0.154346177
#> 19  019 -0.996129798 -1.962203580
#> 20  020 -0.379517960  0.777471825
#> 21  021  0.373719627 -0.896296227
#> 22  022 -0.754985844 -2.092554746
#> 23  023  1.098063439  0.685613085
#> 24  024  0.354677891  0.060380753
#> 25  025 -0.787078496 -0.470535966
#> 26  026 -0.140782215  1.450223310
#> 27  027 -0.298266647 -1.881543278
#> 28  028 -0.001958223  1.520776685
#> 29  029  1.782000954  1.064471442
#> 30  030 -0.228727725  0.546304134
#> 31  031  0.815739761  0.765441023
#> 32  032 -0.910770380 -1.453765481
#> 33  033 -0.353979333 -1.169591269
#> 34  034  1.463538490  0.370457723
#> 35  035 -1.896542019 -1.160019823
#> 36  036  0.905969815  1.065863303
#> 37  037 -1.464148578 -2.270143257
#> 38  038 -1.881431794 -0.443566310
#> 39  039 -0.801650942 -1.228495104
#> 40  040  1.082341527  0.280083855
#> 41  041  0.131563749 -1.214534396
#> 42  042  0.311953081  0.991756484
#> 43  043  0.172610749 -1.094240561
#> 44  044 -0.609241779 -0.906506774
#> 45  045  1.145781738  1.850947649
#> 46  046  0.046644505 -0.273628668
#> 47  047  1.083336902  0.492755815
#> 48  048  1.832560084  0.966088105
#> 49  049 -0.042714941 -0.630789139
#> 50  050  0.414704201 -1.077506807
#> 51  051 -0.847386715 -1.359995002
#> 52  052 -1.252581252 -0.644119919
#> 53  053 -0.261904787  0.560860472
#> 54  054  0.130034642 -0.353568647
#> 55  055 -1.461895805 -1.901650306
#> 56  056  0.593308944  0.979454076
#> 57  057 -0.803626063 -0.308269094
#> 58  058  1.046873364 -0.278258339
#> 59  059 -0.390109764  0.329101348
#> 60  060  0.460943733 -0.646290662
#> 61  061 -1.008607502 -0.851527751
#> 62  062  1.261785997  0.378697150
#> 63  063  0.837406001  0.371078035
#> 64  064 -0.397212305  0.947145545
#> 65  065 -0.159129917  0.523846743
#> 66  066  1.079735581  0.265692560
#> 67  067 -0.677409834 -0.867898074
#> 68  068 -0.385982811  0.155694564
#> 69  069 -0.677760382 -1.938455446
#> 70  070  0.204526403  0.913517368
#> 71  071  0.418198568  0.569310715
#> 72  072 -0.355763316 -0.597870766
#> 73  073 -0.723753045 -1.319898126
#> 74  074 -0.247828160  0.341583158
#> 75  075 -1.019154907 -1.150121182
#> 76  076  1.929010514 -0.133952188
#> 77  077 -1.703559858 -1.028344301
#> 78  078 -0.493456914 -0.001388414
#> 79  079 -0.395137405 -0.871568606
#> 80  080 -0.465071854 -1.624610874
#> 81  081 -0.601686096 -1.081426335
#> 82  082 -0.684726074  0.372925679
#> 83  083  0.207298086 -0.968763677
#> 84  084 -0.458241049  0.548586835
#> 85  085  0.282654676  1.438126201
#> 86  086 -0.033657287  0.589600750
#> 87  087  1.458796978  1.823699553
#> 88  088  0.006366313 -0.275893748
#> 89  089  0.609924740 -0.277637841
#> 90  090 -1.249248703  0.095114842
#> 91  091  3.793236901  2.883326764
#> 92  092  0.314400722  0.450327784
#> 93  093 -0.654211492  0.456644851
#> 94  094  0.359933781 -0.584384252
#> 95  095  0.561058719 -0.567769243
#> 96  096 -1.094625325  0.140894516
#> 97  097  0.425191274  0.317284489
#> 98  098 -1.024521242 -0.254038091
#> 99  099  2.336751096  0.490403947
#> 100 100  0.397248808  1.112792576

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.23222413 -0.337224366
#> 2   002  1.90987509  0.505602300
#> 3   003  1.07835169  1.006548517
#> 4   004 -1.57954609 -0.083403814
#> 5   005  1.00632097 -0.598032587
#> 6   006 -1.71229201 -1.019837609
#> 7   007  0.62081523  1.355000165
#> 8   008  1.21279674  0.548160796
#> 9   009  0.19573455  0.692998427
#> 10  010 -0.24743330  1.340570796
#> 11  011 -0.78869584 -0.145442502
#> 12  012 -0.05347371 -1.121336796
#> 13  013 -2.45512452 -1.619144072
#> 14  014  0.59097975  0.697443037
#> 15  015 -1.33858501  0.636594489
#> 16  016 -0.14173169 -2.006974912
#> 17  017  1.47578732 -0.820975125
#> 18  018 -0.53561487 -1.189845433
#> 19  019 -0.49944012 -0.376250803
#> 20  020  0.28405175  1.411439575
#> 21  021  1.64543603  0.141964424
#> 22  022 -0.90731228 -0.609257982
#> 23  023 -1.13247992 -1.767468340
#> 24  024 -0.57768242 -0.146986766
#> 25  025  1.12800400  0.807786436
#> 26  026  0.40736413  1.143665908
#> 27  027 -1.19156554 -2.425483110
#> 28  028 -0.02364494 -0.418961909
#> 29  029  0.80689669 -0.055963463
#> 30  030 -0.70757314  1.464236609
#> 31  031 -0.92153877 -0.593256890
#> 32  032  1.26464789  0.159925872
#> 33  033  1.30592679  0.346591385
#> 34  034 -1.02280620 -2.266428645
#> 35  035 -1.90296247 -1.229321446
#> 36  036  1.23542909  1.119932517
#> 37  037 -0.37762699 -0.611417951
#> 38  038 -1.33846675 -1.257984462
#> 39  039  0.90024759 -1.051766496
#> 40  040 -0.14007924  0.262820684
#> 41  041 -1.22578019 -1.115671770
#> 42  042  1.22192272  1.382343791
#> 43  043 -1.18028355 -1.274215038
#> 44  044  0.89763038  0.517454354
#> 45  045 -0.61460878  0.391322622
#> 46  046  0.01261002 -0.489677286
#> 47  047  0.32536123  0.658030409
#> 48  048  0.72901494  0.664702990
#> 49  049  1.66647352 -0.115889988
#> 50  050 -0.29165858 -0.068064519
#> 51  051  0.29062279  0.188038228
#> 52  052  0.63535990 -0.883036221
#> 53  053  0.97812671 -0.084568873
#> 54  054 -0.95382801  0.475703623
#> 55  055  0.93502484  0.731809431
#> 56  056  1.83926771  1.405581443
#> 57  057 -0.60616493  0.552472282
#> 58  058  0.60739227  0.952759945
#> 59  059  1.58177142  0.074858053
#> 60  060  1.01558411  1.077357709
#> 61  061  1.46640404  1.021462895
#> 62  062 -0.59917270 -0.124661696
#> 63  063 -0.22965154 -0.578039757
#> 64  064 -1.66138669 -0.322012841
#> 65  065 -0.75337928  0.935829989
#> 66  066  0.47383736  1.017781340
#> 67  067 -0.04726526 -0.829324093
#> 68  068 -0.58181190  0.406743418
#> 69  069  1.06642352  1.289201550
#> 70  070 -0.20524323 -0.923092239
#> 71  071  0.39473684 -0.968886161
#> 72  072 -0.38361299 -0.964538079
#> 73  073  1.16574879  3.410112410
#> 74  074 -0.72967069 -1.188180555
#> 75  075 -0.38669489 -0.666909454
#> 76  076  0.22326167  0.164161628
#> 77  077  2.14399091  2.140357988
#> 78  078  0.44361703 -0.710651128
#> 79  079 -1.13579677  0.533131287
#> 80  080 -1.99377967 -0.894416975
#> 81  081  0.58052480  0.169574703
#> 82  082 -2.15602167 -1.512807465
#> 83  083  0.19537673  0.739806081
#> 84  084  1.30297401  0.861184600
#> 85  085  1.24644192  1.525160156
#> 86  086  0.65327116 -1.080722986
#> 87  087 -0.04418092 -0.678652367
#> 88  088  1.04407285  0.007231178
#> 89  089  1.37139209 -0.111460015
#> 90  090 -0.41071777 -0.339690096
#> 91  091  0.48788856 -0.281141186
#> 92  092  1.85227257  0.923939788
#> 93  093 -1.49199090 -1.412918025
#> 94  094 -2.54932338 -1.918077271
#> 95  095 -0.14605085  2.041862312
#> 96  096  0.17132923  0.707412262
#> 97  097 -1.84022540 -0.567504945
#> 98  098 -0.44738834 -0.542629414
#> 99  099  0.88314154  1.385819736
#> 100 100 -0.49940052 -0.296442509