Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4657092
#> Y_2 0.4657092 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.48996977  0.6256859
#> 2  1.90711083 -0.2949942
#> 3  0.06593702  1.6831796
#> 4 -0.20125109 -0.1031050
#> 5 -0.45655546  0.5669474

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.677171159 -0.09617633
#> 2   002 -1.028654639 -0.75075248
#> 3   003  0.388403819  0.62784139
#> 4   004 -0.450540352 -1.00653998
#> 5   005 -0.031151954  0.42270336
#> 6   006 -0.686535481  0.14076646
#> 7   007 -1.111926149  0.21185713
#> 8   008  0.043566432  0.17252783
#> 9   009  1.311111098  0.47661941
#> 10  010  1.458925505  1.40629897
#> 11  011 -1.628840402  0.05818876
#> 12  012  0.730383948  0.96453581
#> 13  013  0.162760618 -0.32659315
#> 14  014 -0.706218629 -0.89074725
#> 15  015  0.595808705  0.71122160
#> 16  016  0.986372906  1.82195417
#> 17  017 -2.525223270 -0.28188204
#> 18  018  0.135896476 -0.59343813
#> 19  019 -0.758040362 -0.95900698
#> 20  020  0.652458452  0.44083261
#> 21  021  0.940474310  0.15970158
#> 22  022 -1.419617908  0.43254174
#> 23  023  0.270297193 -1.55310181
#> 24  024  1.128793754  2.41103603
#> 25  025 -1.106249035 -0.14947767
#> 26  026 -0.198504665 -0.45820808
#> 27  027 -0.442404171 -1.18694554
#> 28  028 -0.050622849 -0.59092989
#> 29  029  0.303565675  0.33258827
#> 30  030  0.181066692  1.38182788
#> 31  031  0.495303385 -0.46244252
#> 32  032 -0.451972679  0.19624738
#> 33  033  0.310465629  1.07814845
#> 34  034 -0.987899007  0.41927401
#> 35  035 -1.246485076 -0.93872800
#> 36  036  1.905351059  1.37072205
#> 37  037 -1.898595277 -1.78494534
#> 38  038 -0.544001218  0.05757605
#> 39  039  0.164303648  0.85088151
#> 40  040 -0.063300851  0.58259458
#> 41  041 -0.374895585 -0.21026921
#> 42  042 -0.349605876 -0.60363664
#> 43  043  1.201278993  1.17358508
#> 44  044  0.673958550  1.22325159
#> 45  045 -0.098137025 -1.61966757
#> 46  046 -1.054184524  0.42380451
#> 47  047 -0.501669862  0.47225786
#> 48  048 -0.246520460  0.03987999
#> 49  049  0.051080333 -0.30003912
#> 50  050  0.186267011  0.29630981
#> 51  051  0.201909733 -1.09842264
#> 52  052  0.073824672  0.70603662
#> 53  053  0.564944417  0.06603125
#> 54  054 -1.090345452 -0.99992404
#> 55  055 -0.652011166 -0.17167057
#> 56  056 -1.186462046 -1.17668947
#> 57  057 -0.486517587 -0.92591111
#> 58  058  0.240618939 -0.32979437
#> 59  059 -2.577482788 -1.71214177
#> 60  060  0.356882479 -0.21547132
#> 61  061  1.732376677  0.86186283
#> 62  062  0.278820692  1.00958544
#> 63  063  1.119864718  0.74141049
#> 64  064 -0.248939703  0.90694594
#> 65  065 -0.382665485 -1.38818572
#> 66  066 -0.057686336  0.13133924
#> 67  067 -1.022901736 -0.72070534
#> 68  068 -0.708579150 -0.59933087
#> 69  069  0.008052261  1.05816309
#> 70  070  0.796128567  0.84260884
#> 71  071 -0.260094000  0.11735519
#> 72  072  0.802352257 -0.63801993
#> 73  073 -0.361103015 -1.88164141
#> 74  074 -0.044982420 -1.01460599
#> 75  075 -2.385282354 -0.46838714
#> 76  076  0.151697909 -0.59409641
#> 77  077  0.706326843  0.52222974
#> 78  078 -0.258362043  0.16509575
#> 79  079  1.610163882  1.31096469
#> 80  080 -0.559501700 -1.80024141
#> 81  081 -0.789579082 -0.95479935
#> 82  082 -0.114239322  1.30728949
#> 83  083  0.087158518  0.10596760
#> 84  084 -0.229421295  1.99745610
#> 85  085 -0.515074510 -1.14370219
#> 86  086  0.623805773  0.70946488
#> 87  087 -0.858195068  0.33027634
#> 88  088  0.490175721  0.65657511
#> 89  089 -1.439555180 -2.08541811
#> 90  090  0.109645570  0.55627204
#> 91  091  1.544570217  0.06440816
#> 92  092  0.555388750  0.28871688
#> 93  093  1.533447592  0.79444337
#> 94  094 -1.678400843 -1.42985053
#> 95  095  1.257965835  0.88990578
#> 96  096  0.230519669  0.01548785
#> 97  097  1.361530163 -0.91924558
#> 98  098  0.380867894 -0.20908121
#> 99  099  0.952655250 -0.19727435
#> 100 100  1.082359066  1.26265924

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.34525110  0.067183157
#> 2   002  1.62204427  1.108171263
#> 3   003  0.85763931  0.395083814
#> 4   004 -0.96221165 -2.181439026
#> 5   005 -0.16564399  1.273902455
#> 6   006  0.23572672 -0.542202840
#> 7   007  0.86053729 -0.002827564
#> 8   008 -0.21338920  0.456858171
#> 9   009  1.26531376  1.247925225
#> 10  010 -0.64056543  1.363090450
#> 11  011 -0.65180664 -0.177501551
#> 12  012 -1.71863435  2.354936891
#> 13  013 -0.50422700  0.435271019
#> 14  014  1.94028878  0.860871156
#> 15  015 -2.11639574 -0.716024964
#> 16  016 -1.13329656 -0.958175606
#> 17  017 -0.46444268  0.009454096
#> 18  018  0.68843468  0.150746872
#> 19  019 -0.87997023 -0.223316647
#> 20  020 -0.83471676 -0.589484650
#> 21  021 -0.38570731  1.265639685
#> 22  022 -0.21837361 -1.043893830
#> 23  023 -0.02478434 -0.047070835
#> 24  024 -0.80315564 -0.201847176
#> 25  025 -1.94148938 -1.136298543
#> 26  026 -2.59910360 -0.191339980
#> 27  027  0.97680292 -0.152266800
#> 28  028  0.79554778  0.515106975
#> 29  029  0.32497475  0.177608811
#> 30  030  0.54273009  0.412942533
#> 31  031  0.68097699 -0.059170819
#> 32  032 -1.58822366 -2.041165791
#> 33  033 -0.73653609 -0.458125021
#> 34  034  0.26105949 -0.370716686
#> 35  035 -1.44509469 -1.506269813
#> 36  036  0.23051014  1.342251379
#> 37  037 -0.71180457 -0.967641027
#> 38  038 -1.49022521  0.439383607
#> 39  039  0.95668808 -0.318314796
#> 40  040 -1.05943670 -0.218181863
#> 41  041 -1.73612901 -1.676178707
#> 42  042  0.87441275  1.337283410
#> 43  043  0.21720441 -0.242262641
#> 44  044 -0.12987811  0.388155881
#> 45  045 -1.03915801 -0.133633403
#> 46  046  0.36534802 -0.708162239
#> 47  047 -0.43061567 -0.763248178
#> 48  048  1.87022833  1.519589881
#> 49  049 -0.03670346 -0.947138105
#> 50  050 -1.14591712 -0.185306933
#> 51  051 -0.71904053  0.859008733
#> 52  052 -1.20473094  0.351526083
#> 53  053  1.53333383  0.651187593
#> 54  054 -0.29622682  1.173071941
#> 55  055 -0.76886494  0.751654237
#> 56  056  0.02016048 -0.785215916
#> 57  057  0.75724496  0.740453128
#> 58  058 -0.66912601 -0.930034305
#> 59  059  0.79371913 -0.138078029
#> 60  060  0.28416300  1.016301107
#> 61  061 -0.37601979 -0.488217618
#> 62  062  0.55719396  2.001644636
#> 63  063  0.77692623  0.855953365
#> 64  064 -2.35362942 -0.405369715
#> 65  065  1.34720909  0.727699419
#> 66  066 -1.79732370 -0.384132292
#> 67  067 -0.42797282 -1.772482673
#> 68  068  0.38948324  1.638741344
#> 69  069  3.33835567  1.416707746
#> 70  070  0.48208138  0.699364653
#> 71  071 -2.24772739 -1.655857958
#> 72  072  1.96770325  1.513172815
#> 73  073 -1.71281553 -0.883188533
#> 74  074 -0.66412901 -0.477530786
#> 75  075 -0.15067604  0.145837308
#> 76  076 -0.94542536 -0.986011128
#> 77  077  0.07653269  0.902890059
#> 78  078  0.28480376 -0.243699065
#> 79  079 -0.47181354  0.158267181
#> 80  080 -0.02175683  1.324898656
#> 81  081 -0.79317392 -0.168572248
#> 82  082 -2.24065051 -2.363553491
#> 83  083  0.27624938  0.551892745
#> 84  084 -2.11683481 -0.337600319
#> 85  085 -0.35305848 -0.015358276
#> 86  086  0.29103627  0.587216408
#> 87  087 -0.96100254 -0.810445707
#> 88  088 -0.65114131 -0.021173512
#> 89  089  0.65846853  0.358235773
#> 90  090  0.17906937  0.126129667
#> 91  091 -0.47282643 -0.705345734
#> 92  092  0.23187155  0.049207857
#> 93  093 -0.81341223  0.177485890
#> 94  094 -1.63484083 -0.694988633
#> 95  095  0.52358617  0.946319613
#> 96  096 -0.60288874 -0.431319459
#> 97  097  1.14821749  0.724067437
#> 98  098 -0.21537674  0.639599830
#> 99  099  0.20414199 -0.894646598
#> 100 100 -0.59547903  0.377071356