Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4875281
#> Y_2 0.4875281 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.18160627 -0.2030310
#> 2  0.66956436  0.1748093
#> 3 -0.04524512  1.5698117
#> 4 -1.24088140 -1.9819869
#> 5  0.33875496  0.2030937

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.38439662  0.337984484
#> 2   002  0.97740077  0.700767589
#> 3   003 -1.73337585  0.190825449
#> 4   004 -0.38055932 -1.055405968
#> 5   005  0.02458564  0.489802326
#> 6   006 -0.16011672  0.320186630
#> 7   007 -0.25010339 -0.781003263
#> 8   008 -1.40306621 -2.304065915
#> 9   009 -1.84925709 -0.729256892
#> 10  010  0.04356690  1.767791574
#> 11  011 -0.12836297  0.075048589
#> 12  012  0.32240934  1.433290243
#> 13  013  0.12471524 -0.128813838
#> 14  014  0.43037875 -0.617409099
#> 15  015 -0.42426613 -0.404106693
#> 16  016 -0.87470011  0.005301335
#> 17  017  1.42394394  1.291811850
#> 18  018  0.11198894 -0.500153015
#> 19  019  0.59492964  0.364566312
#> 20  020 -0.05070682 -0.139068036
#> 21  021  1.14557885  1.599202694
#> 22  022 -0.74801160 -0.007648807
#> 23  023  0.93909580  2.168619592
#> 24  024 -0.10442694 -0.698232143
#> 25  025 -0.75580475  0.726356068
#> 26  026  0.30000897  0.124133205
#> 27  027  0.95047546 -0.699287453
#> 28  028  0.30192576  1.236279775
#> 29  029  0.10012503 -0.216894067
#> 30  030  0.43961265  1.631021196
#> 31  031 -1.57909414 -2.197853356
#> 32  032 -1.25816434 -0.942670912
#> 33  033 -0.56733253 -0.338589454
#> 34  034  1.20787796  0.041689605
#> 35  035 -0.05101688 -0.648065329
#> 36  036 -0.49408088  0.004611247
#> 37  037 -0.86766961 -0.393256602
#> 38  038  0.60878340  0.044588124
#> 39  039 -0.85288369 -0.092312719
#> 40  040 -1.18895001  0.376530677
#> 41  041  0.04649346 -0.040613088
#> 42  042  0.77818761 -0.254881387
#> 43  043 -1.39452214  0.753952202
#> 44  044 -0.68686567 -0.058324112
#> 45  045  0.81339419  0.823688973
#> 46  046  0.23352901 -0.124030180
#> 47  047 -0.43966835 -1.286849533
#> 48  048  1.37180814  2.190699775
#> 49  049  1.92024647 -0.652579006
#> 50  050 -0.79320555 -0.269680160
#> 51  051  0.38368610  2.153510962
#> 52  052  0.11543836  2.052099572
#> 53  053 -0.16649484 -0.372844521
#> 54  054  1.16592736  2.048050754
#> 55  055  1.21194147  0.401779461
#> 56  056 -1.07383261 -0.840073657
#> 57  057 -0.75842778  0.133543826
#> 58  058  1.21500620 -0.111035916
#> 59  059  0.02914894 -0.177351972
#> 60  060 -0.94448687  0.065295127
#> 61  061 -0.50995895 -1.484460037
#> 62  062 -0.95726380  0.015528292
#> 63  063 -0.72622911  1.476319419
#> 64  064 -0.05285262 -1.301870147
#> 65  065  0.97351791 -0.218640445
#> 66  066  0.12579272 -1.542175009
#> 67  067  1.37724530  0.682882514
#> 68  068  0.69734417  0.301240106
#> 69  069 -0.66870404 -0.532077414
#> 70  070 -1.26597894 -0.375249861
#> 71  071 -1.77332216 -0.237535004
#> 72  072 -0.48318257 -0.226146781
#> 73  073  0.63568745 -1.718357303
#> 74  074  0.44338473  0.154286210
#> 75  075  1.25667926  0.909586842
#> 76  076  1.19943459  0.558789764
#> 77  077  1.49279183  1.934449715
#> 78  078 -1.31647547  0.239259547
#> 79  079  0.26440766  1.168345723
#> 80  080  0.71994255  0.066249814
#> 81  081 -1.05822056 -0.242398790
#> 82  082 -1.77145227 -1.821653279
#> 83  083 -0.83323909 -0.150399726
#> 84  084 -1.16817234 -1.281236560
#> 85  085 -0.38956919  0.072487976
#> 86  086  1.04171013 -0.001068463
#> 87  087 -0.10422849  0.927353620
#> 88  088  0.51211577  0.303791064
#> 89  089  0.43334266  0.521226483
#> 90  090  0.51994519  0.327400539
#> 91  091  0.77191725 -0.003993931
#> 92  092  0.61469255 -0.812283971
#> 93  093 -2.61940065 -1.911075798
#> 94  094  0.48578106  0.228849241
#> 95  095  0.38864954  1.179443796
#> 96  096  1.06505095  1.521884500
#> 97  097  0.06769994  0.257074226
#> 98  098  0.23213297 -1.429552661
#> 99  099  0.86502787  0.821815573
#> 100 100  0.33759746  1.748494152

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.507398702 -0.42130580
#> 2   002  1.710317997  1.30836759
#> 3   003 -1.583672053 -0.14358991
#> 4   004 -0.798357881 -1.59314470
#> 5   005 -0.695159747 -0.99050438
#> 6   006 -0.208909729  1.03559664
#> 7   007  1.507860857 -0.81338233
#> 8   008 -0.922029395 -2.04814535
#> 9   009 -0.934645994 -0.59542297
#> 10  010  2.143958699  1.42078053
#> 11  011  0.116472520 -1.60982597
#> 12  012  0.921643359  0.45626637
#> 13  013 -0.696821783  1.97864772
#> 14  014 -1.044733205 -0.99432624
#> 15  015 -1.276810613 -0.37985240
#> 16  016  0.586656877 -1.19095319
#> 17  017  0.902094031  1.14800837
#> 18  018  0.605767636  1.01332899
#> 19  019  0.383850809 -0.81489836
#> 20  020  1.620780163  0.08971460
#> 21  021  0.086482380  1.41780992
#> 22  022 -0.548216353 -1.41534194
#> 23  023 -0.716458067 -2.06033599
#> 24  024  0.353191590  1.22932244
#> 25  025  1.697672928  1.82789756
#> 26  026  0.714346161  1.73361982
#> 27  027 -0.643038713  0.78690542
#> 28  028  1.738506167  0.15614098
#> 29  029 -0.527911839 -0.09893777
#> 30  030 -1.084932406 -0.02210602
#> 31  031 -0.934864503 -0.58805455
#> 32  032 -0.234371260 -0.14601890
#> 33  033 -0.972539447  1.53792564
#> 34  034  0.290140662  0.41989047
#> 35  035  0.424039046  0.60173417
#> 36  036  1.383860781  1.06753178
#> 37  037  1.119067891  0.61266803
#> 38  038 -1.995918383 -0.78840033
#> 39  039  1.172159703  0.66615903
#> 40  040  2.327026973 -0.85096653
#> 41  041 -1.196265004  0.21258838
#> 42  042  0.231928661 -0.52880178
#> 43  043  0.812092523  0.60511717
#> 44  044  0.492923815  1.17132705
#> 45  045 -0.857453686 -0.89348014
#> 46  046 -1.127514945 -0.48995305
#> 47  047 -0.084874890 -0.58332021
#> 48  048  1.084974994  0.56004642
#> 49  049 -0.573335778 -0.22689031
#> 50  050  0.363167925  0.21953474
#> 51  051 -0.314745729 -0.41008845
#> 52  052 -0.369650886 -0.14817646
#> 53  053  1.402388558  1.33946193
#> 54  054 -0.854160649 -0.03207251
#> 55  055  0.277338949  1.12261591
#> 56  056  2.025927717  1.30931180
#> 57  057  0.380141032  0.25812849
#> 58  058  0.786246469  0.22236021
#> 59  059  0.431548336 -0.23548632
#> 60  060  0.658531519 -1.32314457
#> 61  061  0.151440941 -0.07264448
#> 62  062  0.104896043 -0.32190308
#> 63  063 -0.534719798 -1.32062548
#> 64  064  1.897515791  0.88426162
#> 65  065  1.257914977  0.82058242
#> 66  066 -0.188302649 -0.58917069
#> 67  067 -0.022846854 -0.26210189
#> 68  068  0.949431350 -0.29744647
#> 69  069 -0.939939242 -0.74383427
#> 70  070 -0.908317110 -1.63837233
#> 71  071  0.678334839  0.38332325
#> 72  072  1.542024401 -0.26256270
#> 73  073  0.177228822  1.33217102
#> 74  074  0.700568769  0.90145676
#> 75  075 -0.695818169  0.06894906
#> 76  076  0.127825670  1.19663080
#> 77  077  0.177016112 -0.37720141
#> 78  078  0.896432297  0.24021809
#> 79  079 -0.506494623  2.40351244
#> 80  080 -1.697987337  0.15860546
#> 81  081 -0.667562517  1.48711773
#> 82  082 -0.999680579  0.32143301
#> 83  083 -0.784799773  0.62595236
#> 84  084 -0.568999844 -1.22095464
#> 85  085 -0.178619004 -0.11814245
#> 86  086  1.667988599  1.41427621
#> 87  087  0.083226669  0.35565741
#> 88  088 -1.612903082  0.80202445
#> 89  089  1.780858846  2.22616520
#> 90  090 -0.371061664 -1.95315264
#> 91  091  0.832586869  1.52594087
#> 92  092 -0.941648810 -0.84201607
#> 93  093 -0.695017607 -0.83540449
#> 94  094 -0.694057113  0.61180403
#> 95  095 -0.812526666 -1.65469862
#> 96  096  0.556727798  0.48679354
#> 97  097  1.129793290  0.61576800
#> 98  098 -0.956099000  0.38515753
#> 99  099  0.002348424  0.77629283
#> 100 100  1.609900173  0.58555687