Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4358897
#> Y_2 0.4358897 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.83840798 -0.1092646
#> 2  1.22628445  0.1035217
#> 3  0.83113455  1.2831950
#> 4 -0.26221228 -0.3752293
#> 5 -0.07276806 -0.4683753

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.86045821  0.756980950
#> 2   002  0.15811695 -0.564572062
#> 3   003 -1.07077239 -2.118849285
#> 4   004  0.24944596 -1.429682255
#> 5   005 -0.53927153 -0.251706966
#> 6   006 -0.60858246  0.382782571
#> 7   007  0.11794146  1.642150923
#> 8   008  0.01063026 -0.466438651
#> 9   009 -0.76341961 -0.991597826
#> 10  010  1.97254472  0.574036640
#> 11  011  0.84068203 -0.727432511
#> 12  012  1.11075238  1.808650107
#> 13  013 -0.03722050 -1.431067317
#> 14  014  0.72813569  1.245424231
#> 15  015 -0.69468886  0.532880640
#> 16  016  0.58249498  0.068291405
#> 17  017  1.89379324 -0.191800004
#> 18  018  2.57530063 -0.156285285
#> 19  019 -0.15394621  1.003828471
#> 20  020  0.59279724  0.473433022
#> 21  021  1.08045723  0.898268968
#> 22  022  1.48738623  1.468724078
#> 23  023  0.80938929  1.489459849
#> 24  024 -0.23335500  0.387829463
#> 25  025 -0.56011759 -0.738018611
#> 26  026 -0.21057337 -0.506984908
#> 27  027 -0.28026837  0.177588721
#> 28  028 -0.66978256 -1.034306119
#> 29  029  0.51279475 -0.111860930
#> 30  030 -1.85922331 -2.202038922
#> 31  031 -0.92295669 -1.088608442
#> 32  032  0.87631161  1.602572344
#> 33  033  2.44431892  2.048719762
#> 34  034 -0.92477210  0.010096787
#> 35  035 -1.68963025 -0.010249254
#> 36  036  0.62046970 -0.404794306
#> 37  037 -1.02829173  0.808732950
#> 38  038  1.33733800  1.191499631
#> 39  039 -1.35469556 -1.533365918
#> 40  040  0.56899732  0.228858882
#> 41  041  0.35236793  0.343200638
#> 42  042  0.87117137  0.467119153
#> 43  043 -1.16937963 -1.010641552
#> 44  044  0.46126768 -0.779228446
#> 45  045  1.77286587  1.471261149
#> 46  046  1.07073682  0.272780394
#> 47  047 -0.05954714 -1.369710336
#> 48  048 -0.11892670  0.297005162
#> 49  049 -0.85809683 -0.958454878
#> 50  050  1.32905897  0.507238752
#> 51  051 -1.24730598  1.432147353
#> 52  052 -0.44909946 -0.376453124
#> 53  053 -1.90457698  0.421887813
#> 54  054  0.97363247 -0.776667160
#> 55  055 -0.97613176  1.051945848
#> 56  056 -0.59180209 -0.593848243
#> 57  057  0.65814263 -0.157310793
#> 58  058 -0.58809688 -1.503758912
#> 59  059  0.33753263 -1.000633083
#> 60  060  0.94256997  2.050809636
#> 61  061 -0.18690817  0.462240035
#> 62  062 -1.71174858 -1.394173986
#> 63  063  0.17753006 -2.266635833
#> 64  064  0.07694696 -0.451774495
#> 65  065 -0.52151473  0.767428748
#> 66  066  0.69922716  0.305303362
#> 67  067 -0.04131293 -0.216177466
#> 68  068 -0.46526021  1.250310101
#> 69  069 -1.90365890  1.627709882
#> 70  070 -0.82523590  0.061695307
#> 71  071  0.95428800 -0.047657418
#> 72  072 -0.24117936 -0.006898538
#> 73  073 -0.50472417 -1.647231977
#> 74  074  0.80419737  0.023047910
#> 75  075  1.20663266  1.496955167
#> 76  076 -1.33833126 -1.567920135
#> 77  077  0.63550827 -0.190099664
#> 78  078 -0.82780269  0.128877136
#> 79  079  0.06248986  0.429959379
#> 80  080 -0.97252992 -1.051339331
#> 81  081  2.25685649  2.121708209
#> 82  082  0.88905245  0.867063560
#> 83  083  1.11155797 -0.430564237
#> 84  084 -0.41261840  0.821058298
#> 85  085  0.80403526 -0.095740387
#> 86  086  0.71013913  0.240217868
#> 87  087  1.03398525  0.129837453
#> 88  088 -0.07966938  0.954431603
#> 89  089 -2.60382195 -0.713830574
#> 90  090 -0.57802540 -0.844336480
#> 91  091 -1.88729604 -0.188893189
#> 92  092 -0.96206436 -0.729160995
#> 93  093  0.13573070  1.820702220
#> 94  094 -1.02360170 -1.155060733
#> 95  095  1.12863255 -0.266848847
#> 96  096 -2.02927109 -1.923498715
#> 97  097 -1.69042558 -1.255422687
#> 98  098 -0.07286872 -1.791796560
#> 99  099 -0.47607341 -0.960892451
#> 100 100 -2.54337014 -1.022191236

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.78519129 -1.8019914346
#> 2   002  0.74876004  0.4358282147
#> 3   003 -0.68906526  0.0806682409
#> 4   004 -0.69522484 -0.2562986777
#> 5   005 -1.09547358 -1.2970556496
#> 6   006 -0.64439329  0.4928362013
#> 7   007 -1.15118734 -0.6018032058
#> 8   008  0.13472288  1.3994527276
#> 9   009  0.53721158  0.9818909603
#> 10  010  1.96064952 -0.0004439756
#> 11  011  1.14149326  0.4328047458
#> 12  012 -1.46528393 -2.1689375661
#> 13  013 -0.46891007  1.3363413343
#> 14  014  1.27783078  1.0160934086
#> 15  015 -0.70002960 -1.0943173209
#> 16  016  0.53700286 -0.0438463465
#> 17  017  0.08934305  1.3606676203
#> 18  018  0.58779547 -0.2881231180
#> 19  019 -0.11006275 -0.1782146812
#> 20  020  0.38402360  0.1796996997
#> 21  021 -0.93498992 -1.2844454088
#> 22  022  0.46888178  1.4168101737
#> 23  023  0.19462646  0.6151698769
#> 24  024 -1.36488004  0.4814583099
#> 25  025  0.97188142 -0.2881562674
#> 26  026  0.41057693  2.3136831067
#> 27  027 -3.23706954 -1.5150744968
#> 28  028  0.46333527  0.8111839478
#> 29  029  0.67726464  0.2994350487
#> 30  030 -1.83388676  0.2408473520
#> 31  031  0.36222377  0.0542869819
#> 32  032 -0.72610342  1.2852037553
#> 33  033 -0.14243315  0.0124935054
#> 34  034 -0.88191586 -0.5441255183
#> 35  035  1.22152835  0.6127516118
#> 36  036 -0.36748238 -0.5712106421
#> 37  037  0.53137024  0.6008901238
#> 38  038  0.91788366  0.4988262114
#> 39  039 -1.59265363 -1.3681825570
#> 40  040 -0.22699277  0.2247418328
#> 41  041 -1.18097314 -0.9643407504
#> 42  042  0.91953710 -0.3844697116
#> 43  043 -0.40071336 -1.0458980829
#> 44  044 -1.79960588 -1.5932828756
#> 45  045  0.42993324 -1.2742737679
#> 46  046 -0.46516063 -0.4234098737
#> 47  047 -0.37851502 -0.2436907578
#> 48  048 -0.27795348  0.4356745077
#> 49  049 -0.07277034 -0.6019570750
#> 50  050 -0.13095028  0.4760288656
#> 51  051 -0.05269801 -1.4653990689
#> 52  052  1.69611617  1.1291163054
#> 53  053 -0.67079559 -0.1127241605
#> 54  054 -0.64585050 -1.4185694416
#> 55  055 -0.11925160 -0.8856016463
#> 56  056  0.81142256  0.8355564595
#> 57  057 -0.60934991 -1.7709379362
#> 58  058  0.62392486  0.9496681862
#> 59  059 -2.02591143 -0.9568799848
#> 60  060  0.08106684 -0.2761701402
#> 61  061  1.05414467 -0.3252501267
#> 62  062 -0.63320965  0.3698690775
#> 63  063 -0.49632576 -0.5288892594
#> 64  064  1.35765649  1.2837620848
#> 65  065  2.23228869  0.6774571333
#> 66  066  0.16008552 -0.3255339249
#> 67  067 -1.47024477 -0.2909386486
#> 68  068 -0.35849961  0.1085432928
#> 69  069 -0.37094492  0.0233435729
#> 70  070  1.32141973  1.0236937442
#> 71  071 -0.61815000  0.2846922921
#> 72  072 -1.80276539 -0.2269745364
#> 73  073  0.21745711 -0.1502951741
#> 74  074 -0.70867632  0.0015311022
#> 75  075 -0.50057260 -0.1938558292
#> 76  076 -2.80652338  0.0238695112
#> 77  077 -0.30043357  0.6140731003
#> 78  078  0.43621237 -1.2550861642
#> 79  079  1.44746219  0.4030902500
#> 80  080 -0.73181194 -0.2003368427
#> 81  081  1.25118594 -0.3668958402
#> 82  082 -0.07986341 -0.2590559289
#> 83  083 -1.92060162 -2.0983578560
#> 84  084 -1.61589325 -1.1906924737
#> 85  085  0.06436805  1.0366026527
#> 86  086 -0.64715356 -0.5627423071
#> 87  087  0.60920468 -0.0964887827
#> 88  088 -0.46196244  1.8669743313
#> 89  089 -0.58595744 -0.0483517989
#> 90  090 -0.99199583 -1.5542328463
#> 91  091  0.39397641 -0.0103520207
#> 92  092 -0.87025252 -0.1948320859
#> 93  093 -1.39675012 -0.6883888240
#> 94  094  0.75293519 -0.0473531937
#> 95  095  0.75169534 -0.8149084241
#> 96  096  0.81757609 -0.4713332161
#> 97  097  0.69462756 -0.5754725505
#> 98  098  1.45219302  1.1590513093
#> 99  099 -0.23026947  1.0736519290
#> 100 100  0.87621061 -0.0392927739