Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5059365
#> Y_2 0.5059365 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.7637647  0.13976361
#> 2 -1.3106851 -0.09329371
#> 3  0.1156161 -0.40856954
#> 4  0.1178330 -0.88615702
#> 5 -1.2585147 -2.41078543

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.652928560  0.55791572
#> 2   002 -1.348891594 -1.97938932
#> 3   003  0.174410582 -0.31146419
#> 4   004  0.616183980 -0.61857148
#> 5   005 -0.370880807 -2.04999103
#> 6   006 -0.573361999 -0.03408167
#> 7   007 -0.746759362 -0.48089835
#> 8   008 -1.052457819 -1.04001316
#> 9   009 -0.166669298  0.66536898
#> 10  010  1.341920756 -0.65132531
#> 11  011 -1.249644860 -0.45668676
#> 12  012 -1.807427607 -0.81206504
#> 13  013  0.260818896 -1.36371855
#> 14  014  0.007687821  1.22536043
#> 15  015  0.620080915 -0.50037320
#> 16  016 -0.141042120 -0.18979278
#> 17  017  1.107443280 -0.34229163
#> 18  018  0.366155484  1.40818825
#> 19  019 -0.769401309  0.01169339
#> 20  020 -0.338111244 -1.13998580
#> 21  021  0.903716838 -0.42761996
#> 22  022 -0.844969089 -0.09073573
#> 23  023 -0.458701424 -0.23291700
#> 24  024 -0.835208274 -1.72481850
#> 25  025 -0.994122807 -2.33157586
#> 26  026  1.087189101  0.67069745
#> 27  027 -3.023518454 -1.82610427
#> 28  028  0.059782222 -0.45047533
#> 29  029 -0.775732666 -1.13949712
#> 30  030  0.168819430  0.60775099
#> 31  031  1.974107428  1.03243168
#> 32  032 -0.611484142 -0.77246457
#> 33  033  0.824919986  1.70146451
#> 34  034 -1.015206245 -0.62932988
#> 35  035 -0.571086040 -1.44104094
#> 36  036  1.653797512 -0.29448861
#> 37  037  2.579490821  1.51314309
#> 38  038 -0.710186030 -1.08463416
#> 39  039  1.620072402  0.67163383
#> 40  040  0.008595293 -0.49173608
#> 41  041 -0.171077148  1.20519910
#> 42  042 -0.645746329 -0.30385649
#> 43  043 -0.263906013  0.28981193
#> 44  044  2.153569539  0.52282656
#> 45  045 -0.093186821  0.83044757
#> 46  046 -1.746652217 -0.69254012
#> 47  047 -0.639868603  0.15391814
#> 48  048 -0.078932522 -0.15085991
#> 49  049 -1.632766442 -1.21591736
#> 50  050  1.342910872  0.74610314
#> 51  051 -0.947278793  0.37523959
#> 52  052 -1.184689360 -1.28740160
#> 53  053  0.574611182  0.17463384
#> 54  054  0.446938573 -0.27365312
#> 55  055 -0.701957244 -0.81000796
#> 56  056  0.477187517  1.53056179
#> 57  057 -0.436792311  0.05999146
#> 58  058 -0.885867836 -0.81600731
#> 59  059  1.159408159 -0.25739367
#> 60  060  0.188130560 -0.91272881
#> 61  061  1.030092446  0.71361395
#> 62  062 -0.320785471 -1.13833706
#> 63  063  0.250464363  0.07871777
#> 64  064 -0.693663348 -0.06965966
#> 65  065  0.772345803  1.35318320
#> 66  066 -0.627822609 -0.08072967
#> 67  067  1.644698837  2.15555118
#> 68  068 -0.531487874 -0.08219221
#> 69  069 -1.359015057 -0.53088084
#> 70  070  0.782944873  1.75843938
#> 71  071  1.421082477  1.22531229
#> 72  072 -0.774024451  0.38931494
#> 73  073  0.250716473 -1.20485297
#> 74  074 -0.732130885 -1.64886136
#> 75  075  0.377441510 -0.80365387
#> 76  076  0.894556500  1.13204463
#> 77  077  1.806372824  0.74202524
#> 78  078  0.378162251 -1.38323397
#> 79  079 -0.715359567  0.30646003
#> 80  080 -1.048110057 -2.35947657
#> 81  081  0.852524070  2.11694189
#> 82  082 -1.297381590 -0.33091773
#> 83  083 -0.437578776  0.05716220
#> 84  084 -0.659395618 -0.52186611
#> 85  085 -1.503928042 -2.30828529
#> 86  086 -1.080891533  0.37248300
#> 87  087  1.384492672  1.44822461
#> 88  088  1.229372207  1.19246760
#> 89  089 -0.697416256  0.04186884
#> 90  090 -0.984158939 -1.48645561
#> 91  091  0.904823750  0.65046162
#> 92  092  0.223543556  1.25748527
#> 93  093 -1.464384793 -1.46249965
#> 94  094 -0.347794229 -0.96543256
#> 95  095 -0.639470259  0.08309369
#> 96  096  0.671307056 -0.89492139
#> 97  097  0.601738741  0.17878994
#> 98  098 -0.009389620 -0.64760169
#> 99  099  0.665482346 -0.03775927
#> 100 100 -0.208153585  0.75406842

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001 -0.8047257313 -0.42930674
#> 2   002 -0.6975012851 -1.18092847
#> 3   003  0.0553182261 -1.20323293
#> 4   004 -0.3605500626  0.73166926
#> 5   005 -0.4816410513 -0.59364694
#> 6   006  2.3879308481  1.68045177
#> 7   007  0.1828818649 -1.90516922
#> 8   008  0.0987056766 -0.31638366
#> 9   009 -0.3149783885 -0.80710904
#> 10  010  1.1244640234 -0.40140245
#> 11  011  0.7512873202 -0.49174750
#> 12  012  0.6121889181 -0.21281438
#> 13  013  0.2139611169 -0.85810054
#> 14  014 -0.8411945856  1.09099217
#> 15  015 -0.4223307040  0.38215937
#> 16  016 -0.1126244122  1.78391816
#> 17  017 -0.9576037636 -1.20892958
#> 18  018 -0.7203492000  0.95250804
#> 19  019  0.0239964877 -0.13244707
#> 20  020 -1.2635333125  0.21660247
#> 21  021 -2.2609319732 -0.20656601
#> 22  022  0.7687074305  2.47778822
#> 23  023 -0.3196988027  0.58913502
#> 24  024  0.8355331406  0.30148613
#> 25  025 -1.2423290996  0.90417765
#> 26  026 -1.6794649335 -0.25874234
#> 27  027 -1.1560018390 -3.53429429
#> 28  028  0.2767730081  0.34594014
#> 29  029 -0.5016829294 -0.49017143
#> 30  030  0.7266109461 -0.49015830
#> 31  031 -0.8210694354 -1.49955582
#> 32  032 -0.9122022194 -0.92167823
#> 33  033  0.7022721330  0.50512025
#> 34  034 -1.7221683296 -1.30444480
#> 35  035  2.3089944612  1.53924351
#> 36  036 -1.8599427059 -2.46104999
#> 37  037  0.5412198821  0.58430923
#> 38  038  0.1333443054  1.28330510
#> 39  039  0.6946471963  0.91249080
#> 40  040  0.3031799357 -0.25201485
#> 41  041  0.5243342672 -0.11264887
#> 42  042  0.7520485383  0.89446026
#> 43  043 -1.4790948777 -2.30807550
#> 44  044  2.1477475634  2.03591599
#> 45  045 -1.2109694682 -1.56595663
#> 46  046 -0.6774488098 -1.69988693
#> 47  047  0.8230707008  1.09043470
#> 48  048  0.0024202600 -0.59173113
#> 49  049 -0.6963257312 -1.63335519
#> 50  050  2.5774511472  1.70909462
#> 51  051 -0.0710848712 -0.38438244
#> 52  052  0.2172503435  1.08512272
#> 53  053  0.9266260682  0.85154032
#> 54  054 -1.1843178497 -0.26396226
#> 55  055  0.8129523624  0.90041163
#> 56  056  0.8679495910  0.13204664
#> 57  057 -1.2729358789 -0.76253268
#> 58  058  0.8321841418  0.94097641
#> 59  059  0.1613680788  0.20427351
#> 60  060 -0.7534251255 -0.77950436
#> 61  061 -0.6812931603 -1.46079618
#> 62  062  0.7352547179 -0.57246395
#> 63  063  0.6841796815  1.27258836
#> 64  064  1.2805203568  1.81126236
#> 65  065 -0.4450193084 -0.56425663
#> 66  066  0.7461220651 -0.28281688
#> 67  067 -0.8035390015  0.28756323
#> 68  068 -1.4834343697  0.18067489
#> 69  069  0.4693922209  1.25756205
#> 70  070 -0.8074016764 -1.31257666
#> 71  071 -0.9564248467 -1.94432749
#> 72  072 -0.3640233197 -0.79645391
#> 73  073  1.4556654059  0.36067141
#> 74  074 -0.5454369017 -0.98827404
#> 75  075 -0.8065026541  0.01625063
#> 76  076  1.4006514657  0.78553352
#> 77  077  1.8550074035  0.87576630
#> 78  078 -0.6897747585 -0.60319295
#> 79  079  2.4467010097  0.28050634
#> 80  080  0.6755259394  0.34853789
#> 81  081 -0.1577974848 -1.19867110
#> 82  082  0.0006882668 -0.41573674
#> 83  083  0.0479227449  0.59127606
#> 84  084  0.7441316850  0.73828404
#> 85  085  0.5885281807  0.92751814
#> 86  086  0.1450531637 -0.13869388
#> 87  087  0.2467795078 -0.51150990
#> 88  088 -0.0590596913 -0.87371570
#> 89  089  0.4716309949  1.14837399
#> 90  090 -1.9055184432  0.91651806
#> 91  091 -2.0666584724 -1.38593623
#> 92  092  0.2342472855  0.40504999
#> 93  093  0.8087835116  1.08390711
#> 94  094 -1.8847123690 -0.69617194
#> 95  095  0.7051490477  0.89381170
#> 96  096 -0.3938945365 -0.43594085
#> 97  097 -0.0062871570  0.21928959
#> 98  098 -2.1670508409 -0.70897399
#> 99  099 -1.7621416759 -1.25975667
#> 100 100 -0.7604328806  1.88788007