Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5239322
#> Y_2 0.5239322 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.5504225  0.51165196
#> 2 -0.6956415 -0.05592518
#> 3 -1.9024695 -3.13735287
#> 4 -0.7339879  0.42238512
#> 5  0.4749561 -1.02599903

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.38280109  1.631005239
#> 2   002 -1.45402290 -0.557319169
#> 3   003  1.39639444  0.684466019
#> 4   004  1.59434481  1.877366940
#> 5   005 -0.89626560 -0.235763446
#> 6   006 -2.42176674 -2.081129463
#> 7   007  0.44631190  0.730966416
#> 8   008  1.75449810  2.081818979
#> 9   009  0.49736320  0.298275833
#> 10  010 -0.38264094 -0.485590797
#> 11  011 -1.06827766 -0.760726147
#> 12  012 -0.39007500 -2.200168316
#> 13  013 -1.64763668 -1.798168912
#> 14  014  1.04901126 -0.816549065
#> 15  015 -0.31370967  0.387538043
#> 16  016 -0.01251825  1.597356142
#> 17  017  0.15856541  0.339283877
#> 18  018 -0.31534495  0.609706180
#> 19  019 -0.30380477  1.669888851
#> 20  020 -0.39690040  0.221985376
#> 21  021 -1.39682106 -1.585304495
#> 22  022  0.21813612  1.698859162
#> 23  023 -0.98019151 -1.862542271
#> 24  024  0.38958605 -1.900795654
#> 25  025  0.94634567 -0.624153579
#> 26  026  1.42709641  0.951516409
#> 27  027 -1.41207086  0.276538825
#> 28  028  0.32846758  0.351624463
#> 29  029 -1.24440500 -1.482038816
#> 30  030 -0.02832997 -2.136827611
#> 31  031  0.12797600  0.665924310
#> 32  032 -0.04824643  0.401114774
#> 33  033 -2.13527046 -1.388986757
#> 34  034  0.26814117  0.864830862
#> 35  035 -0.25264923 -0.013507274
#> 36  036  1.20684680  0.153345212
#> 37  037  1.95463493  0.858865894
#> 38  038 -0.02817521 -1.349246075
#> 39  039 -0.26835781 -0.297697426
#> 40  040 -0.08426098 -0.008055952
#> 41  041  0.09334095  0.229532764
#> 42  042  1.29665111  0.435729158
#> 43  043 -1.27811663 -1.763673459
#> 44  044  0.49435306 -0.273511611
#> 45  045  0.84010555  0.873304999
#> 46  046  0.34402231  0.928251122
#> 47  047  0.75810243  1.261430878
#> 48  048 -0.04440386 -0.576168609
#> 49  049 -0.88648598 -0.205609542
#> 50  050  0.08827928  0.274064341
#> 51  051 -1.35615798 -1.201077624
#> 52  052 -0.46128761 -1.385206327
#> 53  053 -1.08030238 -0.596196213
#> 54  054  1.04828431  0.598761251
#> 55  055 -0.16432242 -0.644770775
#> 56  056  0.14802173 -0.259383566
#> 57  057 -0.29087247  1.049658378
#> 58  058  1.39061712  0.924038315
#> 59  059  0.10720453  0.362281556
#> 60  060 -0.11193074  0.220805915
#> 61  061  0.32168829 -0.511113822
#> 62  062  0.09590835 -2.954412225
#> 63  063 -0.44399243 -1.227343174
#> 64  064  0.58318488 -0.338909415
#> 65  065 -0.26298031 -1.444237947
#> 66  066  1.06545713  1.342288894
#> 67  067 -0.07204879 -0.763314060
#> 68  068 -0.58115536 -0.556035187
#> 69  069  2.14658910  2.022598612
#> 70  070  0.25357775  0.762002557
#> 71  071  1.39154238 -0.386177949
#> 72  072  0.14420365  0.556527353
#> 73  073  1.27952986  1.655778657
#> 74  074 -0.47007494 -0.005069644
#> 75  075 -2.07267686  0.345301682
#> 76  076  0.27279000 -0.884962731
#> 77  077 -1.26141475  0.489949392
#> 78  078  0.42969508  0.792131307
#> 79  079  1.67158471  0.738990999
#> 80  080 -0.24558967  0.719736361
#> 81  081  0.41432147  1.547985662
#> 82  082  0.47301963 -0.676206029
#> 83  083 -0.04319331  0.247005890
#> 84  084 -0.46120947  0.766931534
#> 85  085 -0.52313921  0.114237778
#> 86  086 -0.71115750 -0.767613872
#> 87  087  1.06738308  0.744058253
#> 88  088  1.78530831  0.123365127
#> 89  089  1.25867587  0.815772124
#> 90  090  1.55895640  0.287410104
#> 91  091 -0.68928543 -0.751664301
#> 92  092  0.00536793  0.052373360
#> 93  093 -1.28769752  1.583108655
#> 94  094  0.42313057  0.259256369
#> 95  095  0.31134809 -0.502487063
#> 96  096  0.41636746  1.506991811
#> 97  097  0.41697414  0.591335605
#> 98  098 -0.02635389 -0.039488471
#> 99  099  0.50790867 -0.143431286
#> 100 100 -1.83407708 -0.275312280

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.864132674  1.046546330
#> 2   002 -1.595563354 -1.065557077
#> 3   003  0.373040989  1.326819437
#> 4   004 -0.038934237 -1.559347303
#> 5   005  0.526707117  0.998246788
#> 6   006 -0.545558690  0.765941683
#> 7   007 -0.402803963 -1.013973816
#> 8   008  0.868400815  1.082517656
#> 9   009  0.018393420  0.692936654
#> 10  010 -0.189216865 -0.647496111
#> 11  011 -0.040629688  1.666217927
#> 12  012  0.415760204 -0.380501161
#> 13  013  0.050492987 -0.245659418
#> 14  014 -0.183385555  0.234668798
#> 15  015  0.607383909  0.690235842
#> 16  016 -0.981814337 -0.619886408
#> 17  017  0.112875784 -0.765448548
#> 18  018  0.147268729 -0.194645152
#> 19  019 -1.549605006 -1.755748499
#> 20  020  2.757706849  0.214196352
#> 21  021 -1.241510192 -1.001988619
#> 22  022  0.852981459 -0.607558536
#> 23  023  0.254944327 -0.998971794
#> 24  024 -1.899680008 -1.363851764
#> 25  025  0.601776952 -0.062976973
#> 26  026 -0.424190542 -0.386386569
#> 27  027  0.497387525  0.142117252
#> 28  028  0.216143854  1.467835997
#> 29  029  1.468756718  1.305036216
#> 30  030  0.792479394  0.565181645
#> 31  031 -0.028930995 -0.638181643
#> 32  032 -0.591100792  0.947410537
#> 33  033 -0.938414561 -0.109401433
#> 34  034 -1.217525022  0.808896281
#> 35  035  1.815737156  2.200342243
#> 36  036  0.470917298 -0.321729565
#> 37  037 -0.029459034  0.014416195
#> 38  038 -0.318164358 -1.397814420
#> 39  039  0.196364685 -0.264317981
#> 40  040 -0.203869626  0.293673825
#> 41  041 -2.437253675 -0.272563701
#> 42  042 -1.666920128 -0.444624713
#> 43  043  0.362318621  0.343802671
#> 44  044 -0.940898745  0.165184534
#> 45  045  0.164331027  0.459456105
#> 46  046  0.356054352  1.289745810
#> 47  047 -0.872173758 -0.249089981
#> 48  048  1.104918243  0.710543765
#> 49  049 -0.068906208  0.000240317
#> 50  050 -0.040901746  0.076803537
#> 51  051 -0.573364399 -0.589082078
#> 52  052 -0.447785849  0.649151714
#> 53  053 -0.207157126 -1.544010092
#> 54  054  0.268389454  0.077155492
#> 55  055  0.469394960  0.387020483
#> 56  056 -1.037003555 -0.182576457
#> 57  057  1.976839631 -0.769524159
#> 58  058 -0.389709172  0.552520424
#> 59  059 -2.288507834 -1.881044788
#> 60  060 -0.108044331  0.468365659
#> 61  061 -0.876179678  0.239648256
#> 62  062  0.416219975  0.393038686
#> 63  063 -0.751204835 -0.635979734
#> 64  064  1.674853907  2.384668566
#> 65  065 -0.006883714  0.846432421
#> 66  066 -1.705043612  0.032816949
#> 67  067  0.014768360 -0.271882933
#> 68  068 -0.935563114 -0.944677434
#> 69  069  1.424714417  2.506056815
#> 70  070 -1.146452417 -0.383232089
#> 71  071  2.544477955 -0.111922932
#> 72  072 -0.960013966 -0.434972044
#> 73  073  2.360484552  1.570933880
#> 74  074 -0.577560094 -1.463973511
#> 75  075  0.633327032 -1.223924419
#> 76  076 -0.097887114  0.543322580
#> 77  077  2.195056807  1.575768952
#> 78  078 -1.216345752 -2.186960124
#> 79  079 -0.315921101  0.104430248
#> 80  080  0.970853277 -0.155863391
#> 81  081 -2.202698782 -1.881479584
#> 82  082  2.354722378  1.380297034
#> 83  083 -0.354329880 -1.490634687
#> 84  084 -0.110879526 -0.402651896
#> 85  085 -0.945764355 -0.104552825
#> 86  086  1.954186118  0.738962355
#> 87  087 -2.242516555 -0.067032166
#> 88  088  0.285379788  0.813435071
#> 89  089 -0.410440745 -0.695583851
#> 90  090  1.601683781  2.737179823
#> 91  091 -0.200225550  1.892976303
#> 92  092 -0.766616882 -1.322125568
#> 93  093  0.142256195  0.907285846
#> 94  094 -0.269829779 -1.242206132
#> 95  095 -1.053529322 -1.193149787
#> 96  096  0.419179751  0.071418284
#> 97  097 -0.436271648  0.245746866
#> 98  098  0.043056287 -0.083881851
#> 99  099  0.286495418  0.021663204
#> 100 100 -1.187089999  1.141092585