Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5235111
#> Y_2 0.5235111 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.09104663  0.7740274
#> 2  0.15087186 -0.0255562
#> 3 -1.34986689 -1.1646694
#> 4  1.53585790  0.4228746
#> 5 -0.56333432 -0.6996736

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.166687940  0.361152126
#> 2   002  1.467360279  0.767157368
#> 3   003 -0.295465839 -0.892781095
#> 4   004  0.210858538  0.003820573
#> 5   005 -1.100458582 -1.414109599
#> 6   006 -0.180442483  0.457648816
#> 7   007 -0.599587345  1.011445679
#> 8   008 -0.713350923 -0.481118070
#> 9   009 -0.160634763  1.118093731
#> 10  010 -0.760261617  0.200367589
#> 11  011 -0.514378761  0.200281750
#> 12  012  0.840995776 -1.133927829
#> 13  013 -1.127039958 -1.303954431
#> 14  014 -1.691916784 -1.380869847
#> 15  015  0.590300339 -0.136071605
#> 16  016 -0.333202732 -0.663713147
#> 17  017  0.826715032  0.679325213
#> 18  018 -0.601824449 -1.355352672
#> 19  019  1.544451020 -0.164182896
#> 20  020  0.896520131  1.675274488
#> 21  021 -0.242099269  0.481368554
#> 22  022 -2.204032398 -2.149344934
#> 23  023 -0.115170977  0.941807192
#> 24  024  1.575268064  1.525896886
#> 25  025 -0.499254912 -1.508575864
#> 26  026 -1.183665825 -0.921362090
#> 27  027  1.164714273 -0.140562617
#> 28  028  0.202873913  1.235784254
#> 29  029  0.359061667 -0.264769883
#> 30  030  0.562145240 -0.353033217
#> 31  031 -0.765992188 -2.583323871
#> 32  032  0.750222646 -0.412232477
#> 33  033  0.314717829  0.396124232
#> 34  034 -0.888603050 -1.275830987
#> 35  035 -1.499013967 -0.896872261
#> 36  036  1.211399440 -0.296746983
#> 37  037  0.523638778  0.418098492
#> 38  038  0.776280918  0.980494324
#> 39  039 -0.505250740  0.973320471
#> 40  040 -0.432282646 -1.656263869
#> 41  041 -0.486910611  1.810617155
#> 42  042 -0.182336882 -1.244603883
#> 43  043 -0.631637845  0.066244550
#> 44  044  0.854512455  0.883861274
#> 45  045 -0.079723862 -1.303820196
#> 46  046  0.683010119  1.949649542
#> 47  047 -0.929318095  0.153000211
#> 48  048  1.511236143  1.316326024
#> 49  049  1.076019017  1.271905684
#> 50  050  2.290717527  0.556594531
#> 51  051  0.501053153 -0.778250078
#> 52  052  0.769480352  1.051832011
#> 53  053 -0.238176488  0.067740690
#> 54  054  1.057568549 -0.171769987
#> 55  055 -1.317266901 -0.111862479
#> 56  056  0.198618718  0.360432158
#> 57  057 -0.813493423 -1.181250244
#> 58  058 -0.229514052 -0.987475131
#> 59  059 -0.022012287  0.377143621
#> 60  060  0.219023678 -1.929197937
#> 61  061 -1.456330639 -0.694404452
#> 62  062  0.648414713  1.141167424
#> 63  063  0.749549508  0.026985934
#> 64  064  1.050048364  0.505461503
#> 65  065 -0.580808929  0.344476896
#> 66  066  0.225033610 -0.675120270
#> 67  067  0.162464980 -0.741891540
#> 68  068  1.398009456  0.589785436
#> 69  069  1.571017481  1.173091782
#> 70  070  0.963459052 -0.352765961
#> 71  071  1.652421005  1.689867948
#> 72  072 -1.818089583  0.312448220
#> 73  073 -0.488614116 -0.056620796
#> 74  074  0.453262590  0.619168863
#> 75  075  0.601867128 -0.957864732
#> 76  076  1.351533449  0.692743299
#> 77  077  0.004998364  2.004192008
#> 78  078  0.228948547 -1.595129305
#> 79  079  0.192219969 -0.237776519
#> 80  080 -0.095807921  0.001094909
#> 81  081  0.073389404 -0.559122056
#> 82  082  0.824673102 -0.257315200
#> 83  083 -0.277046903  1.289751812
#> 84  084 -0.030022317  0.507138829
#> 85  085  1.174043109 -1.067352725
#> 86  086 -0.475094080 -0.996104972
#> 87  087  0.202132107  0.381406074
#> 88  088  0.196630877  1.644184280
#> 89  089  0.263383586 -1.062860260
#> 90  090 -1.085088225 -0.019561110
#> 91  091  0.057146954  0.631308085
#> 92  092 -1.380643076 -1.255689205
#> 93  093 -0.584057445 -1.224651899
#> 94  094  0.166874906  0.318439018
#> 95  095  0.729690008 -0.460418735
#> 96  096  2.073163903  1.295732549
#> 97  097 -0.133365330 -0.897045177
#> 98  098  0.112399730  0.266022474
#> 99  099  0.184527576 -0.395340837
#> 100 100 -0.462934635 -0.850089265

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.10771251 -1.80505049
#> 2   002  0.26959388  1.56966699
#> 3   003  1.94358780  0.67803251
#> 4   004 -0.19134830 -0.20978527
#> 5   005  0.56567958  0.78664976
#> 6   006 -1.03290799  0.36829170
#> 7   007  0.98166425  2.31924798
#> 8   008  0.23925438  0.63308495
#> 9   009 -0.99691938 -0.74489475
#> 10  010  0.56131441 -0.39645738
#> 11  011  0.75015490  0.02013168
#> 12  012 -2.38912344 -0.31138300
#> 13  013  0.94281835 -1.06919761
#> 14  014 -0.99223050 -0.63595829
#> 15  015  1.11962354  0.88066735
#> 16  016  0.97417545 -0.14352959
#> 17  017 -0.17016814 -0.38472212
#> 18  018 -1.43547992 -0.52444765
#> 19  019  2.42066094 -0.37706462
#> 20  020  0.13185837 -1.40207854
#> 21  021  0.57033716 -0.28677258
#> 22  022 -0.23404671 -0.71939855
#> 23  023 -0.27413453 -0.40016189
#> 24  024 -1.43997325 -1.09316203
#> 25  025  0.73698021  0.82382249
#> 26  026 -1.06895981 -1.54044659
#> 27  027 -0.11800006  0.18882184
#> 28  028 -2.12774038 -0.02651436
#> 29  029 -0.02119599  1.11755166
#> 30  030  0.40492302  0.81921493
#> 31  031 -0.70008959 -1.09653605
#> 32  032 -0.49149133 -0.64902405
#> 33  033  1.28919717  1.06462626
#> 34  034  0.84377102  0.14923078
#> 35  035 -0.23593710 -0.34357713
#> 36  036 -0.48116096  0.41123148
#> 37  037 -0.39010209  0.75832171
#> 38  038 -0.30765240 -1.65499092
#> 39  039  1.29103812 -0.92002420
#> 40  040 -0.03101514 -0.05766836
#> 41  041  1.75299967  1.08571963
#> 42  042  0.95590724  2.10911908
#> 43  043  1.23375383  0.82057012
#> 44  044  0.52438587 -1.38645359
#> 45  045 -1.12914659  0.21587652
#> 46  046  0.42492892 -0.31487827
#> 47  047 -1.32374591 -1.25636201
#> 48  048 -1.15516289  0.21621954
#> 49  049  0.72565984 -2.52544921
#> 50  050 -0.15629678  1.13346897
#> 51  051 -1.11729940  0.77384986
#> 52  052 -1.30618034 -0.74480524
#> 53  053  0.13987195 -1.50963716
#> 54  054  1.33323488  0.78738895
#> 55  055 -1.18363032 -2.39674299
#> 56  056  0.35348578  0.83856795
#> 57  057  0.35453434  0.44443632
#> 58  058 -0.04379739  1.01638631
#> 59  059  0.95550747 -0.03898736
#> 60  060  0.21821934  0.47505927
#> 61  061 -1.25265806 -2.46071084
#> 62  062  0.93969234  0.34348088
#> 63  063  2.07744965  1.33524229
#> 64  064  0.29625871  0.26276783
#> 65  065  0.43260878 -0.95373304
#> 66  066 -1.42499189 -0.66029548
#> 67  067 -0.91819915 -0.69226173
#> 68  068 -0.50547790  1.38033405
#> 69  069 -2.03566885  1.11038797
#> 70  070 -1.64256570 -1.02881538
#> 71  071  1.16977052  1.72700648
#> 72  072 -0.37459217 -0.26463019
#> 73  073 -1.34342655 -0.97653527
#> 74  074  0.83733036  1.02924382
#> 75  075 -1.48607994 -1.59278405
#> 76  076 -1.54279590 -1.60651982
#> 77  077  0.17695902  1.01346111
#> 78  078  0.61682958  0.49951582
#> 79  079 -0.14787120  0.60276569
#> 80  080  1.83141513  1.17745770
#> 81  081 -2.25162412 -0.49596117
#> 82  082  0.71562404  2.12010996
#> 83  083  0.10045616  1.23473646
#> 84  084  0.49519474  0.54173002
#> 85  085 -1.33553116  0.65822160
#> 86  086  0.18212153  0.55506459
#> 87  087  0.01851066  0.10885754
#> 88  088  0.14839617 -0.30783564
#> 89  089 -0.51795469 -0.82116417
#> 90  090 -1.37671299 -0.75350586
#> 91  091  0.47931011 -0.20034677
#> 92  092 -0.04385972 -1.46223413
#> 93  093 -0.89137899  0.03341448
#> 94  094 -0.55408458  0.22075690
#> 95  095 -0.12641620  0.29977900
#> 96  096 -0.59434383  0.42376368
#> 97  097 -0.42301993  0.10674891
#> 98  098  0.06342108  0.17342157
#> 99  099  0.20475787 -0.19526044
#> 100 100 -0.31856644 -0.71520070