Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.498326
#> Y_2 0.498326 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1047581  0.8819287
#> 2  1.2654130  0.7326275
#> 3  0.5617947  0.1308514
#> 4 -0.2517260 -0.8154534
#> 5 -1.6449954 -1.2940630

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -1.502529742 -0.98503127
#> 2   002 -0.014260745 -0.96552247
#> 3   003  0.157620820  0.77811969
#> 4   004 -1.612039472 -0.82152747
#> 5   005 -3.899280582 -1.52461919
#> 6   006  2.354744006  0.75395091
#> 7   007 -0.530233065  0.76624069
#> 8   008  0.095527878  0.19875238
#> 9   009  1.027207117  1.56898176
#> 10  010  0.537715027 -0.87844007
#> 11  011  0.272852142 -0.92660381
#> 12  012  1.408610134  1.33570677
#> 13  013 -0.795835029  1.26029752
#> 14  014  0.997761619  0.33366825
#> 15  015 -0.462395544 -0.87246931
#> 16  016  0.088580952  0.22604065
#> 17  017  0.554853314  0.91504406
#> 18  018  1.360193611  0.60895307
#> 19  019  0.242293211 -0.02575183
#> 20  020 -0.104495278 -1.18140333
#> 21  021  2.252114602  1.64293684
#> 22  022  0.891762947  0.97894713
#> 23  023 -2.126049927 -1.29039431
#> 24  024 -0.562024832 -0.28434981
#> 25  025  0.941170700  0.20999498
#> 26  026 -1.137420371 -0.83601915
#> 27  027  0.031543342 -0.81379727
#> 28  028  0.279245411 -0.16348081
#> 29  029  0.025017017 -1.21969982
#> 30  030  0.544601277  0.12630058
#> 31  031 -0.129488060 -1.21498061
#> 32  032  0.275882639  0.52465126
#> 33  033  1.375666492  2.13936640
#> 34  034  2.002605622  1.67683384
#> 35  035 -0.328379352 -0.58490627
#> 36  036 -1.165660988  0.24716252
#> 37  037  1.082668394 -1.08453443
#> 38  038  0.368260514  0.57213601
#> 39  039 -0.927565855 -0.73322381
#> 40  040 -0.264263816 -1.22579069
#> 41  041  0.072707120  0.46003684
#> 42  042 -0.559035895 -0.25489002
#> 43  043 -1.841609650 -0.35423662
#> 44  044  1.121805993  1.02554508
#> 45  045 -0.492184084  0.58418644
#> 46  046 -1.144943168 -1.04278469
#> 47  047  0.594762429  0.33556810
#> 48  048  1.572329937  1.79475524
#> 49  049 -0.920486635  0.24972770
#> 50  050  0.647506724  1.04839851
#> 51  051 -2.425950350 -1.20301655
#> 52  052 -1.140667073  0.48763116
#> 53  053 -1.347047643  0.95715871
#> 54  054  0.598550012  0.58901614
#> 55  055 -0.270749762 -1.04198160
#> 56  056 -0.101132692 -0.40951999
#> 57  057  0.622559548  0.24536113
#> 58  058 -0.699357025 -1.14236112
#> 59  059  1.022659639  1.38630524
#> 60  060 -0.713115381  0.61616529
#> 61  061 -0.552384504 -0.45993283
#> 62  062  0.532635549  0.63412175
#> 63  063  1.601380885  1.66444225
#> 64  064 -0.149412789 -1.11347443
#> 65  065 -0.896585824 -1.52080646
#> 66  066  0.529706896 -0.33775183
#> 67  067 -0.403829539  0.31294729
#> 68  068 -0.424197910 -0.44540263
#> 69  069 -0.485863355  0.16835219
#> 70  070  0.590391497 -0.13403966
#> 71  071  1.646329439  0.49568688
#> 72  072  0.172016046  1.31473185
#> 73  073 -1.460219421 -0.94394564
#> 74  074  1.460705931  1.36240369
#> 75  075  0.742705987 -0.93220469
#> 76  076  0.706706572  0.34954395
#> 77  077 -0.962936903 -0.46120015
#> 78  078  0.447535732  0.56490890
#> 79  079 -0.862281615 -0.10832626
#> 80  080 -0.253594068  0.12417664
#> 81  081 -0.548530446  0.70576858
#> 82  082 -1.552570754  0.11047115
#> 83  083 -0.966494068 -0.27612263
#> 84  084 -0.480934854 -0.51702685
#> 85  085  0.646037860 -0.17438905
#> 86  086  1.011315099  1.56756807
#> 87  087 -0.790963003  1.06828272
#> 88  088 -0.681776020  0.43803915
#> 89  089  1.315157414  1.40851202
#> 90  090  0.008412819  0.67363159
#> 91  091 -0.275209123 -0.40340219
#> 92  092 -0.015213735  0.31238583
#> 93  093 -1.355969820  0.46453717
#> 94  094  0.006610567 -0.83064583
#> 95  095  0.779399624 -1.08815037
#> 96  096 -1.227919145 -1.18565741
#> 97  097  0.164525624 -0.07129913
#> 98  098 -0.108272992  1.08421955
#> 99  099  0.732407147 -0.72908613
#> 100 100  0.692850796  0.01358076

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -2.04254319 -2.1623522747
#> 2   002  0.32106267  1.2241370251
#> 3   003 -2.13170584 -0.4395113648
#> 4   004 -0.83950329 -0.0099292307
#> 5   005 -1.05649764 -0.8632015624
#> 6   006 -0.55018960 -0.0113017726
#> 7   007 -0.08345475 -1.0510871173
#> 8   008 -1.36509121 -0.1632469432
#> 9   009  0.42561060 -0.1004475793
#> 10  010 -0.30745094 -0.5872765950
#> 11  011 -0.91465026  0.0963891208
#> 12  012 -1.29502682  0.8617092100
#> 13  013  0.13421521  1.0211727829
#> 14  014  0.05091484  0.2448588479
#> 15  015 -1.49314301 -0.2246313756
#> 16  016  0.68187078 -0.0284193487
#> 17  017 -0.82526164 -0.9222700050
#> 18  018 -0.94245265 -1.4620955404
#> 19  019 -1.39120892 -1.6047980860
#> 20  020  0.86880721  0.9112222234
#> 21  021 -0.30577991 -1.4123407103
#> 22  022 -0.65703608 -1.1972424379
#> 23  023  1.14013112  0.1757233608
#> 24  024 -0.02191025 -0.8381996282
#> 25  025 -1.47288257 -1.8785939112
#> 26  026  0.09666790 -0.3910293489
#> 27  027 -0.62699949 -0.5374077824
#> 28  028 -0.25831798  1.2021588858
#> 29  029  0.52688788 -2.1992138320
#> 30  030  1.46119900  0.8466396272
#> 31  031  0.28702627 -0.9360591558
#> 32  032  0.20299088  0.2134940344
#> 33  033 -0.66940723 -0.2360028577
#> 34  034 -1.76815590 -1.4784620218
#> 35  035 -2.43633786 -0.9055733220
#> 36  036 -0.20473507 -0.3391510916
#> 37  037  0.71951480 -0.4829547231
#> 38  038 -0.75822750 -0.1519175526
#> 39  039  0.17313993  0.8989847907
#> 40  040 -0.40802849 -0.0192656546
#> 41  041 -0.41389806 -1.4444118289
#> 42  042  0.14803836 -1.2245192971
#> 43  043 -0.81145497 -1.3249524616
#> 44  044  1.37833845  0.1762012015
#> 45  045  0.44883877 -0.8114362440
#> 46  046  0.44507363 -0.2051960852
#> 47  047 -1.98136092  0.6723248891
#> 48  048  0.31730231 -0.3332229172
#> 49  049 -0.16386059  1.0829352979
#> 50  050 -1.34388111  0.1960557547
#> 51  051 -0.57466984 -0.1903364297
#> 52  052  1.01998703  1.2529432551
#> 53  053  1.06039248  0.9359196759
#> 54  054 -0.21809642 -0.0136909822
#> 55  055  0.45982472  0.9226174855
#> 56  056 -0.94593576  0.6554422850
#> 57  057 -0.44555331  0.0001602074
#> 58  058 -0.07971878 -0.2369105459
#> 59  059 -1.54634220 -0.6378656151
#> 60  060 -0.18648692  0.0995163447
#> 61  061  1.50412760  1.3396215888
#> 62  062 -0.20124485 -0.0857799527
#> 63  063 -0.74113801 -0.3023410316
#> 64  064  0.21894917  0.1743942195
#> 65  065  0.17970201  0.7788742461
#> 66  066 -0.66750530 -1.3553586139
#> 67  067  1.22252071  0.2751094243
#> 68  068  1.31840959  0.9588734563
#> 69  069  0.99686086 -0.1640645176
#> 70  070  0.20930174 -0.7920151239
#> 71  071 -1.94359202  0.0091785610
#> 72  072 -0.35058747 -0.8031822674
#> 73  073 -0.45685533 -1.3893332630
#> 74  074  0.66238840 -0.2041060808
#> 75  075  0.18071837  0.6534855221
#> 76  076  1.46287670  0.1205870376
#> 77  077 -0.51634492 -1.0654119497
#> 78  078 -0.65509312  0.7628487824
#> 79  079  0.63232343 -0.4983502947
#> 80  080 -0.20622851 -0.1512775673
#> 81  081 -1.16372338 -1.1024907147
#> 82  082  0.91829700 -0.0409213865
#> 83  083 -0.43135045  0.7047821556
#> 84  084 -0.72665017  0.2330656915
#> 85  085 -0.39831343  0.2117994002
#> 86  086 -1.24735963 -0.5204325433
#> 87  087 -1.35206355 -0.9011869488
#> 88  088  0.83200658 -1.6601209636
#> 89  089 -0.29394549 -0.5179229221
#> 90  090  0.65337659  0.4686411403
#> 91  091 -1.11888381 -0.9899635514
#> 92  092  0.94476778  0.9541949201
#> 93  093  0.07910464  0.1164330049
#> 94  094  2.18373303  2.7350706255
#> 95  095 -0.85284130 -1.1427130410
#> 96  096  1.13489731  3.2285674494
#> 97  097 -1.59361509 -0.8577061593
#> 98  098  1.35551478  0.8933034504
#> 99  099  0.22509117  0.9614456537
#> 100 100 -0.26277328 -0.6790821538