Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5336453
#> Y_2 0.5336453 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  1.12030972  1.60929737
#> 2  0.67750405  1.88692670
#> 3  0.45820191  0.86760977
#> 4 -0.47410874 -0.15717163
#> 5  0.04349239 -0.01930338

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.50170630 -0.14203408
#> 2   002 -0.40218902  0.52956353
#> 3   003  0.37022347  0.01606485
#> 4   004  0.61294088 -0.18821226
#> 5   005 -1.11989021  0.88155244
#> 6   006  1.36119582  1.73652819
#> 7   007  0.30785487  1.62058659
#> 8   008  1.12288505 -0.08699006
#> 9   009  1.55371954  1.18637709
#> 10  010 -0.92762206 -1.03734784
#> 11  011  0.26572569 -0.77457689
#> 12  012 -1.27279459 -0.64088896
#> 13  013 -0.17124753  0.04028370
#> 14  014 -2.39855754 -2.92118267
#> 15  015 -0.56727562 -0.14743216
#> 16  016  0.16862610  0.14384065
#> 17  017  0.07543447  1.20073997
#> 18  018 -0.39632084 -0.41231882
#> 19  019  0.19738538 -1.27517019
#> 20  020  0.63566697  0.31500994
#> 21  021  0.61288777 -1.06849838
#> 22  022 -1.43243736 -0.66180205
#> 23  023  0.54468986  1.11519810
#> 24  024  0.91608535 -0.70003196
#> 25  025  1.31559307  2.06064498
#> 26  026  1.06931274  0.34168434
#> 27  027 -0.73102473 -0.94194381
#> 28  028  0.51902643 -1.50028664
#> 29  029  1.43719426  0.50808195
#> 30  030  0.93688712  0.92676247
#> 31  031  0.73786792  1.82574108
#> 32  032 -0.08161743 -0.52283937
#> 33  033  0.50778925  0.11626977
#> 34  034 -0.75615237 -0.92085375
#> 35  035 -0.05049259  1.43098866
#> 36  036 -0.72064023 -0.33932963
#> 37  037  0.02378358 -2.66740127
#> 38  038  1.53901202  1.20658352
#> 39  039 -1.64481245 -1.83400462
#> 40  040 -1.56900189  0.17972686
#> 41  041 -0.47588742 -0.60874926
#> 42  042  1.08602715  1.10233073
#> 43  043 -1.19677028 -1.18423312
#> 44  044 -0.16005585  0.13354226
#> 45  045 -0.81578966 -0.10993986
#> 46  046  1.74266825  0.80780236
#> 47  047  1.05957572 -0.14410101
#> 48  048  1.09495773  1.22049263
#> 49  049 -0.33065651 -0.06192867
#> 50  050 -1.55557462  0.05724520
#> 51  051 -0.25386070 -1.93336519
#> 52  052 -1.16503748 -0.07838620
#> 53  053  1.10842306  1.58173518
#> 54  054  1.13997739 -0.80160937
#> 55  055 -0.38763922 -0.46745372
#> 56  056  0.17988827  0.49853078
#> 57  057  1.11547900  1.24749030
#> 58  058  0.04142221  0.55114500
#> 59  059 -0.70966913  0.32950362
#> 60  060 -1.04934553 -0.04045700
#> 61  061  0.85638940 -1.03056535
#> 62  062  0.80962793  1.43958584
#> 63  063 -0.43061350  0.95877982
#> 64  064  0.40311978  0.08952005
#> 65  065 -0.54043677  0.92950714
#> 66  066  0.82529136 -0.04643285
#> 67  067  0.91341290  1.92165814
#> 68  068  0.54863875 -0.37465905
#> 69  069  0.09526480 -1.17305580
#> 70  070 -0.77196279 -0.00800780
#> 71  071  1.17093568  1.71515404
#> 72  072  0.56642007 -0.27406359
#> 73  073  0.43430416 -0.14572721
#> 74  074 -1.12900572 -0.61431426
#> 75  075  0.34485324  0.17308312
#> 76  076  1.00450981  0.61522013
#> 77  077  0.62376022  1.15574387
#> 78  078  0.66941594 -0.74723242
#> 79  079  0.48438800  0.24222111
#> 80  080 -0.59979659 -0.04250516
#> 81  081 -0.28074010 -0.06133651
#> 82  082 -2.15554163 -1.66464425
#> 83  083  1.54882752 -1.68712241
#> 84  084 -1.26717634 -1.49817092
#> 85  085  0.62834690  0.73554015
#> 86  086  1.97001448  0.55396830
#> 87  087  1.36989184  1.65094888
#> 88  088  1.03556893  0.78040041
#> 89  089 -1.30534246 -0.59885292
#> 90  090  1.51540534  1.87340319
#> 91  091 -0.72528664 -1.99669667
#> 92  092 -0.19006331  0.69593407
#> 93  093 -0.06477833  0.26359246
#> 94  094  0.34097770  0.93283132
#> 95  095  1.25756899 -0.87481620
#> 96  096  2.23117284  0.72512583
#> 97  097  0.01084103  0.10580666
#> 98  098  1.66510797  0.12952359
#> 99  099  0.52110836 -0.29704240
#> 100 100  0.46733778 -0.32058413

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.16696004  0.5347772035
#> 2   002  0.45465199 -0.2072240796
#> 3   003 -1.02628060  0.2602631491
#> 4   004  0.47912547 -0.0517762797
#> 5   005 -0.30292579 -0.1171665216
#> 6   006 -0.55823665 -2.1626687623
#> 7   007  0.16751689  1.7853938427
#> 8   008 -0.50612739  0.4983558379
#> 9   009 -0.45334413 -0.6229110557
#> 10  010 -1.10234640 -1.0315861811
#> 11  011  0.28262764  0.7661981351
#> 12  012 -0.15295171 -0.2539491600
#> 13  013  1.83402648  0.9175782703
#> 14  014  0.74652369  1.4483334478
#> 15  015 -0.09873261  0.5865098538
#> 16  016  2.00390687  0.4929743886
#> 17  017  0.23481923  0.6415261466
#> 18  018 -2.00462950 -2.4629084216
#> 19  019 -1.43795819 -0.6695049356
#> 20  020  0.12821356  1.2873046182
#> 21  021 -0.17977081  1.2692345749
#> 22  022 -0.39161843 -0.0757338663
#> 23  023  1.55259048  0.5431134506
#> 24  024  0.83921421  1.1060194724
#> 25  025 -0.18594675  1.6710907097
#> 26  026 -0.99390244 -0.1356870681
#> 27  027  0.82362645  0.1939732785
#> 28  028 -0.30220992 -0.3176774271
#> 29  029 -0.78221350 -1.6966785020
#> 30  030 -2.29144141 -0.6713142874
#> 31  031  0.58028250  1.2622861316
#> 32  032  0.67028071  0.0009429683
#> 33  033 -0.45736997  0.1330099209
#> 34  034  1.51709485 -0.3390888425
#> 35  035  0.32512667  0.9916186564
#> 36  036  0.13135089 -0.8429121498
#> 37  037 -1.87540818 -0.6659604968
#> 38  038 -0.22807807  0.0816084204
#> 39  039 -0.04631641 -1.4942410503
#> 40  040  0.84601516  0.2330086788
#> 41  041 -0.40804467 -0.2447522650
#> 42  042  1.08153882  0.5494466108
#> 43  043  0.98217199  1.8075239491
#> 44  044 -0.37580932 -1.4938111875
#> 45  045  0.24974952 -0.8748376607
#> 46  046 -1.25584534 -1.5247704388
#> 47  047 -0.51019235  0.5508113530
#> 48  048  0.57162004  0.3258916626
#> 49  049  0.61464469 -0.3602804179
#> 50  050  0.22126497  0.2683452150
#> 51  051  0.73743694  0.0977182797
#> 52  052 -0.42842971 -0.1137797659
#> 53  053 -0.27496129  0.1680258904
#> 54  054 -0.66680638 -0.6217292678
#> 55  055 -0.43936527  1.0632214784
#> 56  056 -1.33470091  1.1506991726
#> 57  057  1.21507958  1.0602481187
#> 58  058  0.41440285  0.1015645626
#> 59  059 -0.10176455 -0.0112652772
#> 60  060 -0.01655206  0.8763750323
#> 61  061 -0.33730451  0.2202730209
#> 62  062  1.29541494 -0.2338163560
#> 63  063  0.14247922  0.8574014875
#> 64  064  1.19915312  1.3249667441
#> 65  065  1.61380576  1.8289794666
#> 66  066 -1.79819638 -0.6309692866
#> 67  067  0.47669495 -0.9261625769
#> 68  068 -0.37681182 -0.8498593025
#> 69  069  0.42086463  0.8619569282
#> 70  070 -1.22383313 -1.3198272539
#> 71  071 -0.20113031  0.0397278541
#> 72  072 -2.10698052 -1.1429111863
#> 73  073 -0.54726863 -1.4002537822
#> 74  074  1.19099654  1.2461112773
#> 75  075 -0.76775353  1.0273106365
#> 76  076 -0.31581963 -0.6687670723
#> 77  077  0.37188940 -1.2477743275
#> 78  078  1.55546045  2.3845756087
#> 79  079  0.30405924 -0.9544763635
#> 80  080  0.42069107  0.0632448937
#> 81  081  0.39227810 -0.6456180894
#> 82  082 -1.05877184 -0.8973472573
#> 83  083 -0.66645755 -0.1903679194
#> 84  084 -0.72816667 -0.7491138503
#> 85  085  1.61489419 -0.1086321033
#> 86  086 -1.06738478 -0.9649635443
#> 87  087 -0.51483697 -0.8745512578
#> 88  088  0.48619490 -0.5876515697
#> 89  089  0.03461070  0.7128467682
#> 90  090  0.66538335  0.1846873233
#> 91  091 -1.09025519 -0.7217534660
#> 92  092 -0.01581589  0.1898948804
#> 93  093  0.10564756 -0.3686744935
#> 94  094 -0.26543920 -0.3291390637
#> 95  095 -0.52781646 -0.6432414746
#> 96  096 -0.56105728 -1.6376343265
#> 97  097  1.50699344  1.0556593709
#> 98  098 -0.88824381  0.8576907043
#> 99  099  1.29947376 -0.6856043107
#> 100 100 -1.35818087 -0.9376354229