Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4748215
#> Y_2 0.4748215 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.6624012 -0.8025143
#> 2 -0.8240817  0.9114645
#> 3 -0.8553029 -1.8790219
#> 4 -0.2755185 -1.8582420
#> 5 -1.4267695  0.3477551

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.33167703  1.54832978
#> 2   002 -1.45607897 -0.80288617
#> 3   003  0.14988373 -0.98843442
#> 4   004 -1.00747926 -0.37578517
#> 5   005 -0.05038874 -0.22304897
#> 6   006  0.97518321  0.56167242
#> 7   007 -0.62466697 -0.48517106
#> 8   008 -0.05285339 -0.21791861
#> 9   009  1.37207340  0.21058187
#> 10  010  0.15268582  1.41235514
#> 11  011  0.05401385 -0.13822415
#> 12  012  2.03797238  2.34695837
#> 13  013 -0.07541876  0.33274227
#> 14  014 -0.39606054 -0.62488158
#> 15  015 -1.40162628 -0.83688545
#> 16  016 -0.12763389 -1.28190052
#> 17  017 -0.73476200  0.01589772
#> 18  018 -0.31578687 -0.57294639
#> 19  019  0.01103973 -0.20959791
#> 20  020  0.18512750 -1.09233132
#> 21  021 -0.81601716  0.13281240
#> 22  022 -1.18413589 -0.61990824
#> 23  023  0.27784671  0.13254415
#> 24  024 -0.89428300  1.29137351
#> 25  025  0.21925714 -0.98131716
#> 26  026 -0.49148589  0.55550417
#> 27  027 -1.73375099 -0.43798598
#> 28  028 -0.17234935  0.76645828
#> 29  029  1.27122836  0.93588502
#> 30  030 -0.07624292  0.35468853
#> 31  031 -0.36323592  0.63901718
#> 32  032 -0.06616356  0.02633243
#> 33  033 -1.22726644 -0.75535060
#> 34  034 -0.08562830 -0.87447911
#> 35  035  1.15213136 -0.21036623
#> 36  036 -0.63879595 -0.57736196
#> 37  037  1.31431338  0.15395306
#> 38  038  0.82530074  1.27826289
#> 39  039  1.30448798  0.90051447
#> 40  040  0.84571264  0.38522028
#> 41  041  0.44177755  0.59907610
#> 42  042  0.23415742  0.16524367
#> 43  043 -1.75951926 -0.75230387
#> 44  044 -0.32292627 -0.76877483
#> 45  045  0.23284351 -0.81583635
#> 46  046 -1.28588818  0.71314154
#> 47  047 -0.08888986  1.50665721
#> 48  048  0.24765867  0.27268512
#> 49  049 -1.01004207 -0.33358013
#> 50  050  1.92845708  1.51599394
#> 51  051 -2.09556223 -0.01926480
#> 52  052  1.16881144  1.56634073
#> 53  053  0.84056003  0.38484155
#> 54  054  1.38222767  3.57956711
#> 55  055 -0.43472191 -0.02072760
#> 56  056  0.25028718  0.26892038
#> 57  057  1.06106544 -0.99266795
#> 58  058  1.47836367  0.25194794
#> 59  059 -1.61778838 -2.71037428
#> 60  060 -0.83196025 -1.07450650
#> 61  061  2.34921073  1.30956028
#> 62  062  0.26574968  0.30816426
#> 63  063 -0.45413121 -0.25362667
#> 64  064 -1.95597625 -0.99545902
#> 65  065  0.82656572  0.01154318
#> 66  066  2.49698289  1.31136337
#> 67  067  0.39761993  0.17315555
#> 68  068 -0.84434486 -0.30470366
#> 69  069 -2.13355134 -0.49861102
#> 70  070 -0.30170013  0.19203677
#> 71  071 -0.07606586 -0.23767388
#> 72  072 -0.97394912 -0.22357154
#> 73  073 -0.89942933 -0.06209930
#> 74  074 -0.31124869 -2.22857005
#> 75  075 -0.21386962  0.42549248
#> 76  076  1.39898763  3.13093621
#> 77  077  0.52098362 -0.49139970
#> 78  078 -0.38620112  0.37987107
#> 79  079 -1.19556132 -1.25387745
#> 80  080  0.28074891  0.62108219
#> 81  081 -1.94032774 -0.36217430
#> 82  082 -0.31793445  1.75858049
#> 83  083  0.28176053 -0.01375036
#> 84  084 -0.63772924 -0.17910620
#> 85  085 -1.14435113  1.12231388
#> 86  086  0.18875837  0.85768333
#> 87  087  0.93612473 -0.35486971
#> 88  088 -1.37974641 -1.21498438
#> 89  089 -1.07930602 -0.76047447
#> 90  090 -0.71682342 -1.24962933
#> 91  091  1.68314768  0.60531682
#> 92  092  0.34565960  1.17146297
#> 93  093  0.80001197 -0.04053637
#> 94  094 -0.73978999  1.34828206
#> 95  095 -0.18227248 -0.65253465
#> 96  096  1.11652392  0.38266229
#> 97  097 -0.09391563 -0.39575088
#> 98  098 -1.96787955  0.29486278
#> 99  099  1.57231722 -0.12343260
#> 100 100 -2.18261689 -2.20536885

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  1.423548190  0.137353670
#> 2   002 -0.033180461 -1.424378618
#> 3   003  0.691774960  0.483358457
#> 4   004 -0.845721490 -1.365913988
#> 5   005 -0.056441749  0.177958292
#> 6   006  1.161296679 -0.128056526
#> 7   007  0.790106051  1.760053655
#> 8   008  1.220321265  1.277521568
#> 9   009 -0.151172886 -0.408847891
#> 10  010  1.158268321  0.634542720
#> 11  011  0.603627813 -1.239920748
#> 12  012 -0.306469243 -1.312220311
#> 13  013  0.751916882 -0.267676310
#> 14  014  0.525366353  1.387204120
#> 15  015  0.629453438 -0.618613226
#> 16  016  1.094020193 -1.466120326
#> 17  017 -1.865759840 -0.240482787
#> 18  018 -0.352265527  1.064508693
#> 19  019  0.552913257 -0.871179065
#> 20  020 -0.327048347 -1.057727588
#> 21  021  0.868244863  0.608353644
#> 22  022 -0.927546836  0.007603766
#> 23  023  0.029572184  1.649162791
#> 24  024  0.044274369  0.984320883
#> 25  025  1.132227398  1.113542792
#> 26  026 -0.716109705  0.675884480
#> 27  027  0.121618106  1.321497347
#> 28  028 -0.539563230  0.038453984
#> 29  029  1.918956541 -0.220151159
#> 30  030 -1.118966978  0.910177464
#> 31  031  0.712166026  0.314998516
#> 32  032  0.573255099 -0.598619855
#> 33  033  0.156948036  0.169178459
#> 34  034  0.047400704  0.790855882
#> 35  035  2.176895232 -1.539065469
#> 36  036 -0.604989934 -0.487143130
#> 37  037  1.035399524  1.178583820
#> 38  038  0.294203656  0.681397884
#> 39  039  0.876205342 -0.349475228
#> 40  040  1.867668600  0.891791663
#> 41  041 -0.398542255 -0.731145812
#> 42  042 -0.524823926 -0.304674192
#> 43  043  2.072708343  0.597276168
#> 44  044  1.597911940 -0.008314208
#> 45  045 -0.431553191  2.394369736
#> 46  046 -0.336756665  1.041681201
#> 47  047  1.182122301  0.615516944
#> 48  048 -0.737180700 -0.041530900
#> 49  049  0.420471619 -1.058050431
#> 50  050  1.475322794  1.196086706
#> 51  051 -0.974897823 -0.908762014
#> 52  052  0.127567686 -0.733168964
#> 53  053 -1.222332274 -1.399471611
#> 54  054  0.383192574  0.688937803
#> 55  055  0.127547723  1.500156077
#> 56  056  0.004394037  0.426803112
#> 57  057  0.211156115 -1.380715745
#> 58  058  0.117298493  1.309334648
#> 59  059  0.266201776  1.051119527
#> 60  060 -0.827716947 -0.194988414
#> 61  061 -0.342598613  0.238148792
#> 62  062 -0.609787553 -0.525981121
#> 63  063 -2.547555471 -1.701908526
#> 64  064 -2.571624360 -3.285178412
#> 65  065 -0.960138837 -0.110907511
#> 66  066 -0.248310880  0.724651174
#> 67  067  1.851464764 -0.241455095
#> 68  068 -1.735591728 -2.024874380
#> 69  069 -0.708120408  0.848157736
#> 70  070 -0.426091568  0.249032215
#> 71  071 -1.621659701  0.351013003
#> 72  072 -0.282816210  0.080150369
#> 73  073 -0.031761988  1.306213965
#> 74  074 -0.901122218 -0.921267614
#> 75  075 -0.530721285 -0.695798254
#> 76  076  1.165402749  1.256836167
#> 77  077 -1.041717627 -0.132576273
#> 78  078 -0.038779384  0.013054958
#> 79  079  1.167337893  0.687295877
#> 80  080  0.300405413 -0.097007820
#> 81  081  1.329301814  0.621453156
#> 82  082 -0.764854184 -0.265883985
#> 83  083  1.534293910 -0.106015265
#> 84  084  0.991820306 -1.274001980
#> 85  085  0.168588330  0.418412464
#> 86  086 -0.776623555 -1.109583030
#> 87  087 -0.828680467 -1.033359044
#> 88  088  1.059129595  0.986249523
#> 89  089  1.453949307  1.775922338
#> 90  090  0.641687662 -0.211877769
#> 91  091  0.192888806 -0.740663482
#> 92  092 -0.465773082  0.119450387
#> 93  093  0.141052550 -1.219400143
#> 94  094  0.704549340  2.135528596
#> 95  095  2.546970651  2.217238711
#> 96  096  0.861985659  0.093609504
#> 97  097  0.896673796  1.308747675
#> 98  098  0.089329463  0.745719826
#> 99  099 -0.034548598  1.255298947
#> 100 100  0.582084767 -0.274637988