Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5114754
#> Y_2 0.5114754 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.0145760 -2.0026754
#> 2 -0.6551791 -0.5331648
#> 3  0.3991373 -1.0986634
#> 4  0.4775707  0.5158182
#> 5 -0.1269500 -0.7867242

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.26993332  0.01937727
#> 2   002 -0.68156727 -0.47955169
#> 3   003 -0.62696350  0.53033420
#> 4   004  0.61357245 -1.15785723
#> 5   005 -1.15252651 -2.65427617
#> 6   006  1.24792188  0.19457237
#> 7   007  1.01346875 -0.41929538
#> 8   008 -0.49923238 -0.09802386
#> 9   009  0.24655747 -1.46817125
#> 10  010  0.21407615 -0.22351470
#> 11  011  0.72193217 -0.20061469
#> 12  012  0.25996995  1.36861847
#> 13  013 -1.30807151 -0.94888987
#> 14  014 -0.46696752 -0.42724091
#> 15  015 -1.23555551 -0.36493923
#> 16  016  1.31481014  1.49715690
#> 17  017 -0.61742323 -1.15183951
#> 18  018 -1.67404609 -0.35694567
#> 19  019 -0.06176394 -0.16477704
#> 20  020 -0.29033577 -0.70947867
#> 21  021  0.82406663 -1.03105189
#> 22  022  0.29239417  0.56953535
#> 23  023  1.06613635  0.51574411
#> 24  024 -0.13695261 -1.40880957
#> 25  025 -0.00823601 -0.16052199
#> 26  026 -0.37557472  1.54324344
#> 27  027 -0.95862982 -0.71633176
#> 28  028 -0.25217751  0.63748653
#> 29  029  1.25788056  1.93129344
#> 30  030 -1.21667299 -1.15569215
#> 31  031  1.56337741  1.02714273
#> 32  032 -0.56590097 -0.04230692
#> 33  033 -2.83537809 -1.32432666
#> 34  034 -1.69913180 -0.75095820
#> 35  035  2.48344070  1.40202551
#> 36  036  1.95398125  1.05788413
#> 37  037 -0.34396881  0.94340529
#> 38  038  1.02177733  1.02772383
#> 39  039  0.38547652  0.24884915
#> 40  040  1.43045124  0.55063671
#> 41  041 -1.05016598 -0.32861767
#> 42  042  0.86896868 -0.33892225
#> 43  043 -1.48361011 -0.64394819
#> 44  044 -0.92352581  0.53713405
#> 45  045  1.67438903 -0.02195481
#> 46  046  0.69027401  0.56906420
#> 47  047  0.72633074  0.84218366
#> 48  048  1.80172771  1.04027613
#> 49  049  1.09990496  0.38710747
#> 50  050  0.35030921 -0.66803822
#> 51  051 -0.68758117  0.84560302
#> 52  052  0.24221976  0.17530095
#> 53  053 -2.34888051 -0.89644062
#> 54  054  0.75772522 -1.26351510
#> 55  055 -0.06940806  0.48019006
#> 56  056 -0.87887725 -0.10232943
#> 57  057 -0.63116634 -0.02723199
#> 58  058 -0.24046501  0.10326821
#> 59  059 -2.64935044 -1.74800939
#> 60  060 -0.07222366 -1.60661897
#> 61  061  0.56733543  1.11418142
#> 62  062  1.37227915  0.49207904
#> 63  063  0.61173357  0.99591441
#> 64  064 -0.58161153 -0.40207102
#> 65  065 -0.29303079  0.90796955
#> 66  066 -1.37378303 -0.12712200
#> 67  067 -0.06041276 -0.97261732
#> 68  068  1.33233345  0.11296241
#> 69  069 -0.87390416 -2.00544545
#> 70  070  0.65032308  0.22135689
#> 71  071  0.03891648 -0.22483510
#> 72  072 -0.08630771 -0.16524219
#> 73  073  1.71849973  1.12636288
#> 74  074 -0.65688012 -0.76372009
#> 75  075  0.32206746  0.16055356
#> 76  076 -1.97971059  0.57543241
#> 77  077  0.59390447 -1.24685734
#> 78  078 -1.57186749 -0.84282401
#> 79  079  1.58456010  1.49856878
#> 80  080 -0.18930930 -0.65673365
#> 81  081 -1.21549593 -0.32555758
#> 82  082  0.84831656  0.25912752
#> 83  083 -1.19862201  0.28330078
#> 84  084 -0.86556594 -0.10330834
#> 85  085  0.65182647  2.29961832
#> 86  086 -0.04953136 -0.58182582
#> 87  087 -0.26755170 -0.06495805
#> 88  088  0.90692866  0.51908027
#> 89  089 -0.62022758 -0.55841202
#> 90  090  1.09738859  0.04925715
#> 91  091 -0.94551372 -1.37502869
#> 92  092 -1.28145593 -1.61871063
#> 93  093 -1.13489320  0.81856071
#> 94  094 -0.52175134 -0.15054910
#> 95  095 -0.51508043 -0.25606247
#> 96  096  1.50325467 -0.19388992
#> 97  097  1.76345524  0.78688617
#> 98  098 -0.04219031 -0.16673718
#> 99  099 -0.46258756 -0.37013460
#> 100 100  0.08208920  0.68911796

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.170464713 -0.40892016
#> 2   002 -0.505934416  1.17228362
#> 3   003 -1.761724415 -0.65037212
#> 4   004 -0.978117910 -1.48170414
#> 5   005 -0.209427735 -0.80570412
#> 6   006 -0.134930786  0.12518292
#> 7   007  1.481037442  0.42374343
#> 8   008  0.281510842  1.35120824
#> 9   009 -0.073305883  0.02144372
#> 10  010 -1.113723019 -0.82821324
#> 11  011 -1.613273583 -1.97845013
#> 12  012 -0.258073888 -1.92487289
#> 13  013 -0.938151866  0.72449611
#> 14  014 -0.582413173 -0.43094620
#> 15  015  0.629452813  1.46754713
#> 16  016  0.537398548  1.47000019
#> 17  017 -0.253307817 -1.40106532
#> 18  018  1.787365196  0.11572778
#> 19  019  1.241746710 -0.06101704
#> 20  020 -1.295432078 -1.67481389
#> 21  021  0.181262796  1.03417876
#> 22  022  1.832260957  1.87844068
#> 23  023  0.509928255  0.28339992
#> 24  024  1.484229104  0.93974599
#> 25  025 -0.317168481 -0.69326268
#> 26  026 -0.123710814 -0.35665054
#> 27  027  0.054555306 -0.10140271
#> 28  028  0.082344042 -1.51848651
#> 29  029 -1.260779508  0.18171775
#> 30  030  0.610820629 -1.20249526
#> 31  031  0.855166344  1.00367377
#> 32  032  1.509978985  1.27350255
#> 33  033 -0.477582081  1.26370471
#> 34  034 -1.154009739  0.59775652
#> 35  035 -0.634119662 -0.27039044
#> 36  036 -0.624607910  0.26405066
#> 37  037  0.109988429 -0.21771129
#> 38  038  0.011790085 -1.67659263
#> 39  039 -0.126214036  0.03229558
#> 40  040 -2.314186510 -0.19939313
#> 41  041 -0.249124910  0.88401714
#> 42  042 -0.985149875 -1.95688384
#> 43  043  0.348775210  0.41844834
#> 44  044 -1.557486867 -1.60859591
#> 45  045 -0.622499724  0.14662717
#> 46  046 -0.817940380  0.42968760
#> 47  047  0.129596105  0.97567512
#> 48  048 -0.682339292 -1.46338020
#> 49  049  0.951106199 -0.80995887
#> 50  050 -1.017340730 -0.56444573
#> 51  051  0.714699647  0.17057725
#> 52  052  0.801932016  1.07176989
#> 53  053  0.285919636 -0.61268084
#> 54  054  0.994105686 -0.08400165
#> 55  055 -0.029676526 -0.91859946
#> 56  056 -1.229233497 -1.36543623
#> 57  057 -0.183486011  0.12394958
#> 58  058  0.971323685  1.25249325
#> 59  059  0.160804468 -0.01015298
#> 60  060  0.121360315 -1.71333433
#> 61  061 -0.167543270 -0.11041743
#> 62  062  1.424043403  0.92242429
#> 63  063  0.010271142  0.33480020
#> 64  064 -1.113519500 -0.92371607
#> 65  065  0.521110430 -0.48862275
#> 66  066  0.487835948  1.30978376
#> 67  067  0.710418193  0.66277830
#> 68  068  1.323537821 -0.18820881
#> 69  069  0.399150094  0.73147506
#> 70  070  1.262748320  0.08593811
#> 71  071 -0.346436852 -0.58050919
#> 72  072  0.174040559  0.04804006
#> 73  073  0.124667443 -1.43829946
#> 74  074  2.360105171  0.36061353
#> 75  075 -0.985543174 -0.07409342
#> 76  076  1.028912751  0.78090872
#> 77  077 -0.006193363  0.61090225
#> 78  078 -0.398593592  0.07984918
#> 79  079  1.090638425  1.20609637
#> 80  080  0.303708191 -0.88697679
#> 81  081 -1.476640409  0.23194801
#> 82  082 -0.907140937  1.47769954
#> 83  083  0.489237445 -0.71772929
#> 84  084  0.702562011 -0.35800355
#> 85  085 -2.025896814  0.06781342
#> 86  086 -0.413189026 -0.29459271
#> 87  087  0.444738191 -0.58826880
#> 88  088 -0.660499496 -0.44629363
#> 89  089 -0.682429262 -0.85560895
#> 90  090  0.052653657  1.10549393
#> 91  091  1.532180664 -0.04543681
#> 92  092  0.924334436  0.37745098
#> 93  093 -0.425606360  0.82558903
#> 94  094  0.249832639 -0.84672363
#> 95  095 -0.369973512 -0.70159101
#> 96  096  0.497309266 -0.30080427
#> 97  097 -0.041228981 -0.23082382
#> 98  098  0.170893618 -0.02255949
#> 99  099 -0.431093842  0.54668471
#> 100 100  0.087485593 -0.77157003