Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4598549
#> Y_2 0.4598549 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.3517022 -0.1320050
#> 2  0.8667622  1.4182893
#> 3 -0.2258060  0.3432173
#> 4 -1.1067238  0.4295900
#> 5  1.6114822  2.3146410

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.55695663  0.36324749
#> 2   002 -0.97945505 -0.10653643
#> 3   003 -1.07129183 -0.31070013
#> 4   004  0.43651611  1.06625303
#> 5   005 -0.02893436  0.36531518
#> 6   006 -0.59016539 -1.24363589
#> 7   007  0.16804889 -0.70433470
#> 8   008 -0.11847540 -0.11039546
#> 9   009  0.80578550  0.34470493
#> 10  010  0.06992933 -0.02158936
#> 11  011  0.12312261  0.68710471
#> 12  012 -0.72463004  0.60967268
#> 13  013  0.09273969 -1.04031935
#> 14  014  1.43048590  0.17455469
#> 15  015 -0.46484736  2.52673091
#> 16  016  0.60061383  1.15329599
#> 17  017  0.65466139  0.81738364
#> 18  018  0.08517858 -0.16955451
#> 19  019 -0.97648083  0.25630228
#> 20  020 -0.75638204  1.22687919
#> 21  021  0.77915514 -1.44185719
#> 22  022 -0.86048318 -0.33790849
#> 23  023 -0.18749044  0.33000574
#> 24  024 -0.47304369 -1.02703880
#> 25  025 -0.86283704 -0.09840974
#> 26  026  0.84970376  0.70526316
#> 27  027  1.33709955 -1.41497022
#> 28  028 -0.50856049  1.75951565
#> 29  029 -0.27531725 -0.10562293
#> 30  030  0.30046718 -0.03714663
#> 31  031 -0.78560959  0.15089003
#> 32  032  1.21289514  0.46086165
#> 33  033 -1.59032037 -2.21324363
#> 34  034  0.29783526 -1.78433684
#> 35  035 -0.35164733 -0.37632415
#> 36  036  0.45550701 -2.45962075
#> 37  037 -0.67063361 -0.77529769
#> 38  038  1.21914023  0.13974515
#> 39  039 -0.04154623 -1.86435935
#> 40  040 -0.37053834 -1.66263598
#> 41  041  0.76559358  0.99990179
#> 42  042 -2.18361296 -2.13448653
#> 43  043 -2.14720124 -1.74153093
#> 44  044 -0.42121564  1.28572700
#> 45  045 -0.48848103 -0.29898350
#> 46  046  1.65675268  0.04980602
#> 47  047 -0.01160930 -0.44004587
#> 48  048 -0.26200139 -0.71030228
#> 49  049  1.25697045  1.32517757
#> 50  050  0.27785882 -0.37904886
#> 51  051  0.11379502  0.02103213
#> 52  052  0.57040236  0.74094122
#> 53  053  1.10832929  0.23965448
#> 54  054 -1.55793609 -2.01873428
#> 55  055  0.09608678  1.34283275
#> 56  056 -0.86529775 -0.20305364
#> 57  057 -1.24106733 -0.64455293
#> 58  058  1.57494085 -0.41343117
#> 59  059 -0.85888380 -1.47261581
#> 60  060 -0.81304516  0.65633547
#> 61  061 -1.90170631 -1.70878973
#> 62  062  0.36524381  1.06102127
#> 63  063 -0.82775441 -0.30394730
#> 64  064  0.47629265 -1.21248189
#> 65  065  1.47897807 -0.91482112
#> 66  066 -0.12198380  0.01227440
#> 67  067 -0.23244122 -0.55675469
#> 68  068  0.25492422 -0.04689204
#> 69  069  0.82644618 -1.07073498
#> 70  070 -0.60334069  0.68666762
#> 71  071  0.27353012  0.19736897
#> 72  072 -1.71817437  0.69821120
#> 73  073 -0.31349178  0.54332433
#> 74  074  0.17824249  0.50325375
#> 75  075  1.04511205  0.17607468
#> 76  076 -0.52924748  0.44867962
#> 77  077  1.21813913  0.36712097
#> 78  078  0.59979382  0.66899420
#> 79  079 -1.66957052 -0.72373201
#> 80  080  1.84813093  1.01458030
#> 81  081  1.30173206  0.52985062
#> 82  082 -0.61369202 -0.02527289
#> 83  083 -0.77428808  1.09354871
#> 84  084 -1.89736318 -1.68642995
#> 85  085 -0.61400267 -1.59758193
#> 86  086 -1.91522466 -1.87064187
#> 87  087  1.25401124  0.13702590
#> 88  088  0.67607863  2.04294605
#> 89  089  0.80770363 -0.96398747
#> 90  090  0.55342869 -0.49548296
#> 91  091  0.06501188 -0.21512441
#> 92  092 -0.14458333 -0.70091290
#> 93  093 -1.36005464 -1.20554299
#> 94  094 -0.67664480  0.33869201
#> 95  095 -0.30171294 -0.04386011
#> 96  096  1.58705682  0.17420459
#> 97  097  0.61005416  0.88114729
#> 98  098  0.39366252 -0.36896575
#> 99  099 -1.04855945 -0.17016903
#> 100 100 -0.47410800  1.29643608

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001 -4.812751e-01  0.25514878
#> 2   002 -7.607184e-01 -1.53154348
#> 3   003 -3.062300e-01 -0.06170803
#> 4   004  1.055548e-01 -0.10368608
#> 5   005  1.955978e-01 -0.24392488
#> 6   006  1.546452e+00  0.74628583
#> 7   007  1.268665e+00  0.77783256
#> 8   008 -2.662954e+00 -2.91313302
#> 9   009  1.497732e+00 -0.28569128
#> 10  010  2.262375e-01 -0.45730116
#> 11  011  5.754592e-01  0.08902458
#> 12  012  1.754284e+00  0.33535498
#> 13  013  2.767224e-01  0.08990805
#> 14  014  9.450499e-01  1.40666415
#> 15  015  1.631412e+00  0.79222035
#> 16  016  1.948038e+00  0.73290233
#> 17  017 -3.400172e-01 -0.39086247
#> 18  018 -1.108318e+00 -1.92776876
#> 19  019 -7.801309e-01 -0.34723263
#> 20  020  6.530172e-01  1.13173873
#> 21  021  2.339313e-01 -1.30467201
#> 22  022 -8.235028e-01  0.14178834
#> 23  023  1.613124e+00  1.92371308
#> 24  024 -7.853947e-02  0.85205298
#> 25  025 -2.789280e-01  0.01149410
#> 26  026 -1.850671e-01  0.04632325
#> 27  027  1.015667e+00  0.76584479
#> 28  028 -3.257289e-01 -1.92936180
#> 29  029  7.105955e-01  1.38605333
#> 30  030  2.232373e-01 -0.22800397
#> 31  031 -1.078529e+00  0.25869608
#> 32  032  7.988660e-01  0.60528606
#> 33  033  1.470694e-01  0.20325976
#> 34  034 -5.262109e-01  1.77970593
#> 35  035  2.105440e+00 -0.66353155
#> 36  036 -2.026825e+00 -1.71953783
#> 37  037 -6.183449e-05  0.93135203
#> 38  038  6.524760e-01 -0.31900381
#> 39  039  5.939054e-01  0.65530452
#> 40  040  1.948458e+00 -1.02771479
#> 41  041  1.977483e-01  1.31911991
#> 42  042  1.245679e+00  1.07176763
#> 43  043  6.721110e-02 -0.64174235
#> 44  044  6.779022e-01  0.46307211
#> 45  045 -1.235391e-01 -0.99774305
#> 46  046  1.950445e-01  1.52576648
#> 47  047 -1.384519e+00 -1.27738948
#> 48  048  1.938945e-01  0.25774168
#> 49  049 -2.775830e-01 -0.49414180
#> 50  050 -2.275703e-01 -0.37936699
#> 51  051  2.107184e-01  0.77322268
#> 52  052 -6.926329e-01 -0.65582561
#> 53  053 -7.078560e-01  0.41199805
#> 54  054 -5.087209e-02 -0.51573912
#> 55  055 -1.145898e+00 -0.52914537
#> 56  056 -6.575600e-01  0.38803611
#> 57  057 -3.508013e-02 -0.18733129
#> 58  058  8.493694e-01  0.51665690
#> 59  059 -2.743846e-01  0.99662322
#> 60  060  7.490781e-01  1.15906460
#> 61  061  1.373837e-01 -0.42733305
#> 62  062 -4.352887e-01 -0.53597071
#> 63  063 -6.862796e-01 -0.25861858
#> 64  064 -2.813654e-01 -1.12873122
#> 65  065  1.434350e+00  0.96251499
#> 66  066 -4.910450e-01 -0.07025250
#> 67  067  1.121911e+00  0.16644702
#> 68  068 -4.376825e-01 -0.41589713
#> 69  069  6.614128e-01  0.41872211
#> 70  070 -2.903095e+00 -1.88387382
#> 71  071  2.970221e-01 -0.03904477
#> 72  072 -1.143080e+00 -0.68496091
#> 73  073  1.478854e-01  0.40858723
#> 74  074 -2.178388e-01  0.19183233
#> 75  075  1.642434e+00  2.50379174
#> 76  076 -6.187959e-01 -0.12889340
#> 77  077  1.137191e+00  0.35032026
#> 78  078  6.988580e-01  0.53991964
#> 79  079 -8.718529e-02  1.73848774
#> 80  080  6.494413e-01 -0.83853175
#> 81  081  1.703727e+00  0.70012163
#> 82  082  1.101553e+00  0.51247657
#> 83  083  2.580211e-01 -0.07411277
#> 84  084 -1.042058e+00  0.25740691
#> 85  085 -8.779877e-01 -0.43022092
#> 86  086  1.402790e+00  1.12515223
#> 87  087  2.032173e+00  1.89353324
#> 88  088  4.774928e-01 -0.34625310
#> 89  089 -7.039776e-01  0.18426826
#> 90  090  1.764480e+00  2.10337158
#> 91  091 -2.083687e+00 -1.69331060
#> 92  092  6.272977e-01 -0.22900248
#> 93  093 -3.504683e-01 -0.21774256
#> 94  094 -1.345178e+00 -0.50221702
#> 95  095  3.031354e-01 -0.27078062
#> 96  096  6.587222e-01  2.12523727
#> 97  097  3.964410e-01  0.77023617
#> 98  098  1.139229e+00  0.46744572
#> 99  099 -7.361257e-01 -2.13107966
#> 100 100 -9.688683e-01  0.67393655