Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4898712
#> Y_2 0.4898712 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.0205583 -0.55841790
#> 2  0.3419442  0.04200435
#> 3  0.2164115  0.18205769
#> 4 -0.6080584  0.70259344
#> 5 -0.2130878 -0.39432166

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.455393377 -0.122766845
#> 2   002 -1.794762106 -0.793256126
#> 3   003 -0.625818820 -1.763510512
#> 4   004 -1.000620191 -0.637736509
#> 5   005  0.440693219  1.805731288
#> 6   006 -1.328211213 -0.223194030
#> 7   007 -1.391756845 -1.742700868
#> 8   008  0.106152079 -1.053225009
#> 9   009 -0.153571851  0.526388473
#> 10  010 -2.564153666 -1.231743285
#> 11  011 -1.248173517 -2.471372598
#> 12  012 -0.677234805  0.017085832
#> 13  013 -0.252220926  0.986485634
#> 14  014 -1.451440029 -1.285446278
#> 15  015  0.052406530  0.544648301
#> 16  016 -0.984151195 -2.242469694
#> 17  017  0.903159916  0.644585302
#> 18  018 -0.999641018 -1.490034908
#> 19  019 -0.439422713 -0.709670723
#> 20  020 -2.399934770 -2.113667389
#> 21  021  0.937037927  1.257372589
#> 22  022  0.820153969  0.730711297
#> 23  023  0.376141633 -0.108789935
#> 24  024  0.345469417 -0.855972224
#> 25  025 -2.339183436 -1.371615528
#> 26  026  0.848931002  0.971520771
#> 27  027 -0.875294013 -0.597640962
#> 28  028 -0.149056380  0.864223487
#> 29  029  0.176290842 -0.679083127
#> 30  030 -0.069804989  0.906494333
#> 31  031 -1.088993673 -0.310073524
#> 32  032 -1.354852393 -0.908302459
#> 33  033 -1.730752222  0.795507445
#> 34  034 -0.390432105  0.344702199
#> 35  035 -0.003468279 -0.316082562
#> 36  036 -0.560547902  1.416361219
#> 37  037 -0.218078360  1.227846577
#> 38  038 -0.771221474 -1.002491300
#> 39  039 -0.322762496  1.274799398
#> 40  040  2.286294165  2.596141026
#> 41  041  1.188259388  1.738036090
#> 42  042 -0.178921228 -0.274019655
#> 43  043  0.583744730 -0.145925584
#> 44  044 -0.161856550  0.681780383
#> 45  045 -0.863608710 -2.113351054
#> 46  046  1.370564486  1.478569979
#> 47  047 -0.312976599  0.740532234
#> 48  048 -0.824633217  0.346321347
#> 49  049  0.473669067 -1.519142150
#> 50  050  1.615707258  0.311356929
#> 51  051 -0.406521581 -0.616970268
#> 52  052 -0.060079986 -0.325486509
#> 53  053  0.669706966  0.084012525
#> 54  054  1.551205790  0.363078204
#> 55  055 -1.830410791 -0.745466530
#> 56  056  0.298055643 -0.410619282
#> 57  057 -0.968832714 -0.290138529
#> 58  058  0.164453853  0.738240442
#> 59  059 -1.175078687 -0.026950450
#> 60  060 -0.276357773  0.868513465
#> 61  061  0.299363973  1.662189339
#> 62  062  0.678484662  1.376286315
#> 63  063  1.324680353  1.040943475
#> 64  064  0.231912404 -0.009427278
#> 65  065  0.361739411  0.102733381
#> 66  066  0.512950912 -1.401978745
#> 67  067 -0.130641147  0.058872732
#> 68  068 -0.592769271  0.002962671
#> 69  069  0.519935402  0.680588773
#> 70  070 -0.081925781 -0.292330995
#> 71  071 -0.749069449 -0.327416667
#> 72  072 -0.718770472  0.504417696
#> 73  073 -0.169279100 -0.418314654
#> 74  074 -0.609225027 -0.764674850
#> 75  075  0.645587777 -0.331906055
#> 76  076  0.343604988  0.416212319
#> 77  077 -0.394038710  0.668580021
#> 78  078  0.756761059  0.663256027
#> 79  079 -1.676395780 -1.164764363
#> 80  080  0.251470539  0.189241232
#> 81  081  0.771625807  1.280250110
#> 82  082  0.221047887  1.218275072
#> 83  083  0.009815720  0.476843672
#> 84  084  0.246050974 -0.557367420
#> 85  085  2.258151437  0.357108076
#> 86  086 -0.733486367 -0.814093417
#> 87  087 -0.702846224 -0.842915371
#> 88  088 -0.430851459 -0.136259544
#> 89  089 -1.270090694 -1.045752159
#> 90  090  1.035700248  0.191292494
#> 91  091  0.798668789  0.453661096
#> 92  092  1.609696803  1.017561072
#> 93  093  1.782683510  1.897290392
#> 94  094 -0.982898933 -1.286858859
#> 95  095  2.017245728  1.732529004
#> 96  096  0.784769482  0.722639622
#> 97  097 -0.132576777 -1.007131896
#> 98  098  1.624471214  1.391289260
#> 99  099  2.615573284  2.755138116
#> 100 100  2.437590629  1.616226531

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1           Y.2
#> 1   001 -0.365527221 -0.6285352202
#> 2   002  0.918818898 -1.2028009559
#> 3   003  1.289892042 -0.4492437713
#> 4   004 -0.581562246 -0.7622223035
#> 5   005  1.240762917  0.4369225762
#> 6   006 -1.515190416  0.4574744159
#> 7   007  0.824965364  0.6404135183
#> 8   008  0.002273344 -0.0174779461
#> 9   009  0.809192516  0.1061247537
#> 10  010 -0.762342525 -0.6775886209
#> 11  011 -0.099414337  0.0103166726
#> 12  012 -0.197368254 -0.6763923556
#> 13  013 -1.069575742 -0.7986469395
#> 14  014  0.252032111  0.1759266520
#> 15  015 -0.129420644 -0.3862107632
#> 16  016 -0.041103372  0.4248786873
#> 17  017  0.645190134  0.2602074400
#> 18  018  0.373254556 -0.2784983174
#> 19  019  1.698332695 -0.2429677357
#> 20  020 -0.653665437 -0.2346856560
#> 21  021  0.297573367 -0.1874170078
#> 22  022 -0.133526102  0.3151156559
#> 23  023 -0.711445481 -0.7635167831
#> 24  024 -0.133633901 -0.1097794124
#> 25  025 -0.851234802 -0.9380764583
#> 26  026  1.575560086 -0.4312886050
#> 27  027  0.758954437  1.3861034885
#> 28  028 -0.221866758 -0.6311906333
#> 29  029  1.325865131 -0.2282970950
#> 30  030  2.808281580  2.5852033515
#> 31  031  0.884780474 -0.8953117074
#> 32  032  0.473754064 -0.4850282886
#> 33  033  0.099745941 -0.8294737961
#> 34  034  1.188095690  1.9967420940
#> 35  035 -1.465670546 -0.2480542033
#> 36  036  0.294513226 -1.3350109685
#> 37  037 -0.553844530 -1.0797619809
#> 38  038 -0.449286111 -2.2013108875
#> 39  039  1.137889231 -1.0887976269
#> 40  040 -1.303063878  0.1960586975
#> 41  041 -0.426543469 -0.8000511156
#> 42  042  1.449870448  2.4271128982
#> 43  043 -1.319420189  0.2594090223
#> 44  044  1.725662863  2.4673270547
#> 45  045 -1.272866470 -2.0010357196
#> 46  046  0.037193665  0.4890563009
#> 47  047  1.101194384  1.4570652457
#> 48  048  0.226308297  1.8120024481
#> 49  049  0.009750056 -0.0006361561
#> 50  050 -0.170207278 -0.2100016875
#> 51  051  1.211633674  0.6850880396
#> 52  052 -0.294886103 -0.2793179424
#> 53  053 -0.274001504  1.3929510817
#> 54  054  0.190768011  0.0878160644
#> 55  055 -0.447444408  0.6132046138
#> 56  056 -0.868818892 -1.4197161971
#> 57  057  0.459828081 -0.8747439605
#> 58  058 -0.767042532 -0.6414580397
#> 59  059  1.243781666  1.5276289509
#> 60  060  0.498916292  0.4029112938
#> 61  061  0.538273943  0.6096603644
#> 62  062 -0.508141817 -0.7983625293
#> 63  063  0.292591047  1.5044575600
#> 64  064  2.183168207  1.3250543413
#> 65  065  0.195082237  0.1883549475
#> 66  066  1.912650757  1.2112879382
#> 67  067 -1.694777252 -3.3186871313
#> 68  068  0.922561693  1.1864871862
#> 69  069  0.274486171 -0.3149859728
#> 70  070  0.291224499 -1.7146375942
#> 71  071 -0.454986095  1.8682571325
#> 72  072 -0.070130923 -0.5164507843
#> 73  073 -0.103969455  0.0811146741
#> 74  074 -0.327981614  0.0402184898
#> 75  075 -0.422247350 -0.6831483794
#> 76  076 -0.473559189 -1.6616464073
#> 77  077  0.735967037  0.9977536921
#> 78  078 -0.272132880  0.0965321139
#> 79  079  0.896130580 -1.3698547841
#> 80  080  0.071611467 -0.0602948034
#> 81  081  0.792290200  1.9488510669
#> 82  082 -1.683288629 -0.8022422436
#> 83  083  0.246670398  0.3091434420
#> 84  084 -1.966950588 -2.3745521897
#> 85  085 -0.868987533  0.1643924858
#> 86  086  0.301701319  1.4558420736
#> 87  087  0.605493834 -0.2923894960
#> 88  088  0.458430344  0.4846773697
#> 89  089  1.735275507  2.4745006717
#> 90  090 -0.754677025 -1.5082728310
#> 91  091 -1.490635381 -0.3924311702
#> 92  092 -0.466556819 -0.4305891379
#> 93  093  0.804187110 -0.5379148943
#> 94  094 -0.708304430 -1.4264288481
#> 95  095 -0.845082287  0.2361564362
#> 96  096  1.216727288  2.8450430023
#> 97  097  0.624509768 -0.6565148291
#> 98  098 -0.629618004 -0.6040837684
#> 99  099  1.464559680  1.2199275328
#> 100 100  0.867514581 -0.3726118662