Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5630736
#> Y_2 0.5630736 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.7539853 -1.55931822
#> 2 -0.3052160  0.02423216
#> 3 -0.3916424  0.13890524
#> 4 -3.0988882 -2.12752486
#> 5 -0.4921278 -1.16150609

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  1.441226824  0.49967192
#> 2   002 -0.465191710 -1.14056635
#> 3   003 -0.239319862  1.22248847
#> 4   004  0.545611479 -0.10700675
#> 5   005  0.772289603  0.24758723
#> 6   006  0.310392299 -0.44254450
#> 7   007  0.281778626 -0.48992823
#> 8   008  0.222081662  1.17457031
#> 9   009 -0.332104438 -0.37172331
#> 10  010  0.448393162  0.56720434
#> 11  011  0.257976778  1.07936239
#> 12  012  1.061409545  0.41827757
#> 13  013  0.384982681 -0.03331785
#> 14  014 -1.041717251 -0.82748053
#> 15  015 -1.157414421 -2.80833463
#> 16  016 -1.159137730 -1.64919959
#> 17  017  0.571717726 -0.32277714
#> 18  018 -0.049327601  0.89380165
#> 19  019  1.356450373  0.71795128
#> 20  020 -0.210833491 -0.40161976
#> 21  021  1.365101189  0.10857581
#> 22  022 -0.435038230 -1.18384009
#> 23  023  0.454901086 -0.04514923
#> 24  024 -0.542631755 -0.50166070
#> 25  025 -0.200583840  0.10330971
#> 26  026 -0.589922332  0.17762510
#> 27  027 -0.113013775  0.12508078
#> 28  028  0.327093583 -0.01553204
#> 29  029 -0.632578323 -0.78192939
#> 30  030  0.147642147  1.05423338
#> 31  031 -0.327947246  0.93926233
#> 32  032  2.030156245  0.89775109
#> 33  033 -2.244322398 -2.03540755
#> 34  034  0.091744799  2.57649204
#> 35  035 -0.001919315 -0.71460243
#> 36  036 -1.139673836 -0.75129628
#> 37  037 -0.858908102  0.14103472
#> 38  038  0.624818724  0.42237933
#> 39  039 -0.568050568 -0.09522115
#> 40  040 -0.328143148  0.23014523
#> 41  041  0.790899074 -0.35822700
#> 42  042  1.624751473  1.00391152
#> 43  043 -0.599408869  0.33293007
#> 44  044  0.727830261 -0.44038634
#> 45  045 -0.192216622  0.63815552
#> 46  046  0.450241231  0.70294356
#> 47  047  0.334511700 -0.11590003
#> 48  048  0.608239615  0.42444867
#> 49  049  0.183365577 -0.10679543
#> 50  050 -0.465817684 -1.08142851
#> 51  051  0.228385134  0.32131943
#> 52  052  0.589389059 -1.45311294
#> 53  053  0.479762056  0.74716403
#> 54  054  0.994140601  1.06730541
#> 55  055 -0.511872512  0.96729833
#> 56  056 -1.542056660  0.11505191
#> 57  057 -0.685090928  0.95990540
#> 58  058 -1.433559943 -0.94361915
#> 59  059  1.261475947  0.97512289
#> 60  060 -0.073966093  0.04065383
#> 61  061 -0.334185481  1.23103418
#> 62  062  1.586887255  1.68392376
#> 63  063  0.243036660 -0.41198730
#> 64  064  1.481048984  1.65156793
#> 65  065 -0.451093144 -1.64640543
#> 66  066 -2.020688615 -0.23407976
#> 67  067 -1.847060035 -1.05596967
#> 68  068 -0.713649597  1.15813349
#> 69  069 -0.087223030  0.12251267
#> 70  070  1.085859908 -0.95279634
#> 71  071  1.818892632  0.81633853
#> 72  072  0.228320372 -0.24212535
#> 73  073  0.321325223 -0.32695347
#> 74  074  1.216635439 -0.18190502
#> 75  075 -0.343208348 -0.11839078
#> 76  076 -0.423647638 -0.66137217
#> 77  077  0.876394711  0.62991339
#> 78  078 -0.355391109  1.52853478
#> 79  079 -1.293550477 -1.60643414
#> 80  080  1.121049482 -0.53644746
#> 81  081 -0.212856597 -1.48566333
#> 82  082  0.044707195 -0.20382333
#> 83  083 -0.219046027 -0.92339022
#> 84  084  0.439714941  0.95657992
#> 85  085  0.463671014  0.25036416
#> 86  086  1.474315675  0.55980827
#> 87  087 -1.426766320 -1.08034184
#> 88  088 -1.317628123 -1.05327629
#> 89  089 -0.436820643  0.04881045
#> 90  090 -0.301209597  0.74127759
#> 91  091 -1.005473619 -0.30239911
#> 92  092  0.998330857  0.74487295
#> 93  093  1.038103333  0.10064512
#> 94  094  0.299132940  0.89048593
#> 95  095 -0.696470403 -0.81519556
#> 96  096  1.655651649  0.82929828
#> 97  097  1.128056869  0.80544119
#> 98  098 -0.776621141 -1.04401823
#> 99  099  0.841146116 -0.54525643
#> 100 100  0.496524344 -0.52768912

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.334146199  0.37581039
#> 2   002 -1.263713886 -2.27061280
#> 3   003  1.062505548  1.53309026
#> 4   004 -1.114246614 -1.49199224
#> 5   005 -1.249916225 -0.49786945
#> 6   006 -0.893210368  0.07391657
#> 7   007  0.165977310  1.29820682
#> 8   008  1.108624112  0.61349077
#> 9   009 -0.350875396  0.12700348
#> 10  010  0.631279919  0.23203491
#> 11  011 -0.328712215 -1.14812393
#> 12  012  0.046157383 -1.22661425
#> 13  013  0.052663217  0.47928060
#> 14  014  0.249931201  0.15809573
#> 15  015  0.525295246  0.96951422
#> 16  016  1.783726494  2.28357583
#> 17  017 -1.824484335 -1.56440272
#> 18  018 -0.867026774 -1.18656806
#> 19  019  0.250705222 -0.61540248
#> 20  020 -0.068362773  0.08157083
#> 21  021  0.954602642  0.53814215
#> 22  022  0.237350333 -0.66311191
#> 23  023  0.880938127 -0.38755745
#> 24  024 -0.415473019 -1.04656953
#> 25  025 -0.030497138 -0.12169922
#> 26  026 -2.475487365 -0.12989361
#> 27  027  0.081493173 -1.78485210
#> 28  028 -0.043600513  0.78788568
#> 29  029 -0.327329480 -0.38778656
#> 30  030  1.787102176 -0.58268981
#> 31  031 -0.721258008 -1.97795840
#> 32  032 -0.147875720 -0.23524406
#> 33  033 -0.698537873 -0.67326249
#> 34  034 -0.539458103 -0.62148004
#> 35  035  0.497000609  0.93169355
#> 36  036  0.025922063 -0.18801192
#> 37  037  0.332386868 -0.22590609
#> 38  038  0.840532941 -0.33510641
#> 39  039  1.500373948  0.00487739
#> 40  040 -0.818500560 -1.65423211
#> 41  041 -0.888924798 -1.16800463
#> 42  042  0.076336808 -0.40091948
#> 43  043 -1.252813301 -1.07803256
#> 44  044 -2.077147154 -1.10413156
#> 45  045  0.503883634  1.18469302
#> 46  046  0.241274584 -0.24022999
#> 47  047  0.911098311 -1.34267919
#> 48  048 -0.151535580 -0.89623647
#> 49  049 -0.717510266  1.37438351
#> 50  050 -1.819724179 -0.85321388
#> 51  051  0.009199605  0.22877465
#> 52  052 -1.078846620 -0.66523558
#> 53  053  0.454173355  0.24724553
#> 54  054 -1.213201619  0.68392875
#> 55  055  0.792867785 -1.34199704
#> 56  056 -0.318498630  1.06338274
#> 57  057  1.577257007 -1.10565265
#> 58  058 -0.525269413 -1.17022261
#> 59  059  0.398741257  0.35021782
#> 60  060 -0.493369906  1.49668968
#> 61  061 -1.314636226 -1.33897868
#> 62  062 -0.392877690  1.06112136
#> 63  063  0.111612043 -0.94097142
#> 64  064  0.824411466  1.30357964
#> 65  065 -1.279753327 -0.90915054
#> 66  066  0.570380791  0.45818778
#> 67  067  1.805544137  1.14109547
#> 68  068  1.208357446  0.97052754
#> 69  069 -0.679456369  0.06033177
#> 70  070 -1.113972206  0.57705612
#> 71  071  0.984585215  0.06679000
#> 72  072  0.762848640  2.24481935
#> 73  073  0.636173528 -0.50567610
#> 74  074  2.113923631  2.58501058
#> 75  075 -0.372531730 -0.08714338
#> 76  076 -0.769857906 -0.55540212
#> 77  077 -0.485406689  0.42266801
#> 78  078  0.704998754 -0.11089140
#> 79  079  1.914808983  2.30162969
#> 80  080  1.389485221  1.51287357
#> 81  081 -0.577087920  0.76205249
#> 82  082 -0.308995629 -0.19521140
#> 83  083 -0.479131500  1.17800244
#> 84  084 -0.743572257 -1.50960625
#> 85  085  1.329637290  1.60923323
#> 86  086  0.227749991 -0.52466408
#> 87  087  0.896800528  0.65342268
#> 88  088 -0.736067075 -1.47584529
#> 89  089 -0.959255411 -0.82736593
#> 90  090  2.523695080  1.26343639
#> 91  091 -0.782300139  0.18123351
#> 92  092  0.791091564  0.03791839
#> 93  093  1.262001037  2.35563778
#> 94  094  0.609541740 -0.83362995
#> 95  095 -3.301508702 -1.71427811
#> 96  096  1.758711074  0.10297821
#> 97  097  0.070704639  0.58976243
#> 98  098  0.071537508  0.84310830
#> 99  099 -0.082276477  0.01009644
#> 100 100  1.231972439  0.29750652