Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4556286
#> Y_2 0.4556286 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4860768  0.9216879
#> 2  0.0285559  0.3922260
#> 3 -0.1847095  0.4992808
#> 4 -0.3647319  0.7013196
#> 5 -1.0332949 -0.3742546

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.13752085  0.121360512
#> 2   002  0.19992125  0.812928546
#> 3   003 -0.44196576 -0.274948182
#> 4   004  0.43860979  0.063507796
#> 5   005 -0.41928121 -0.665392518
#> 6   006  0.32373460  0.195950865
#> 7   007 -0.05812740  1.531765856
#> 8   008  0.37023722 -0.774552969
#> 9   009 -1.00038238 -0.956291271
#> 10  010 -0.38313708  0.395272574
#> 11  011 -0.14547227  2.031819843
#> 12  012  1.64881126 -0.891803304
#> 13  013 -0.34450266 -1.222526211
#> 14  014 -0.34865288 -0.953329860
#> 15  015  1.18888153  0.466959009
#> 16  016 -0.99388917 -1.742034216
#> 17  017  0.35412124  1.145688421
#> 18  018  1.80870121  0.804773180
#> 19  019  0.42495018 -1.335577830
#> 20  020 -2.28641525 -2.958845922
#> 21  021  1.42114373  0.607426561
#> 22  022 -0.77468588  0.006714448
#> 23  023  2.43754757  1.122762813
#> 24  024 -0.25833771 -0.867799542
#> 25  025 -1.14955706 -0.319650023
#> 26  026  1.40956969  2.263579733
#> 27  027 -0.21103556 -0.555168718
#> 28  028  0.08856438 -2.112039716
#> 29  029  0.91064952  1.346757716
#> 30  030  0.09735427  0.843165478
#> 31  031 -0.85611786  0.233155330
#> 32  032 -0.81564538 -0.354769410
#> 33  033 -0.44743584 -0.400301403
#> 34  034 -1.52826937 -0.258823236
#> 35  035  0.34892072 -0.649795591
#> 36  036  1.32106415  1.456235638
#> 37  037 -0.49694014 -1.082882607
#> 38  038  1.64344180  0.787160300
#> 39  039 -1.78693939 -0.923802096
#> 40  040 -0.96678258 -0.466892347
#> 41  041 -0.98590153 -0.596361048
#> 42  042  1.52963670  0.853920109
#> 43  043  0.69740361  0.894828497
#> 44  044  1.52382365  1.584685737
#> 45  045 -1.31309226  0.330699723
#> 46  046  1.50376155  2.856998175
#> 47  047  1.83960478  1.322372236
#> 48  048 -0.16703966  0.848008866
#> 49  049  0.77442559  2.428686357
#> 50  050  0.13906023 -1.744289130
#> 51  051 -0.70506180 -0.037031134
#> 52  052  0.27679667  0.729221143
#> 53  053 -0.82270826 -0.890912318
#> 54  054  0.25124289  0.097864830
#> 55  055 -0.82008947 -0.251158722
#> 56  056 -0.38800025 -1.252108321
#> 57  057  0.41644675  0.894896297
#> 58  058 -1.49281321 -0.998095930
#> 59  059 -1.39906714 -0.724777140
#> 60  060  0.54213670  0.040149953
#> 61  061 -0.40692638 -0.138551251
#> 62  062  0.74199163 -2.919509994
#> 63  063 -0.36993289 -1.394815868
#> 64  064  1.10582348 -1.155120273
#> 65  065 -1.27296741 -0.980455340
#> 66  066  0.27654858 -0.228950473
#> 67  067  0.05171648  1.487185528
#> 68  068  0.23151459 -0.376552124
#> 69  069 -0.57708792 -0.843495214
#> 70  070  1.82128184 -0.042960891
#> 71  071  0.74393476  0.732121616
#> 72  072  0.40970033  0.544831646
#> 73  073  0.30893146 -0.516594113
#> 74  074  0.23168335  0.920152160
#> 75  075  0.36438767 -0.280723874
#> 76  076 -1.21113310 -1.589543349
#> 77  077 -0.02261552  0.296890156
#> 78  078  1.72873668  0.595504290
#> 79  079  2.42951582  0.309121906
#> 80  080  0.61587433  0.765002354
#> 81  081  1.09615926 -0.010519812
#> 82  082 -0.17650177 -0.618147511
#> 83  083  0.25067244 -0.199709865
#> 84  084 -0.38430326  0.049453280
#> 85  085 -0.61419428 -1.164146657
#> 86  086  0.62620412  1.432760541
#> 87  087  0.58013029  0.958957347
#> 88  088 -0.70479102  0.869877967
#> 89  089 -0.29056055 -0.683955243
#> 90  090 -1.68990437 -2.298213528
#> 91  091  3.09391878  0.131291694
#> 92  092  0.55590872  1.724881914
#> 93  093 -0.27771729 -1.433115515
#> 94  094 -0.79598248 -0.844481794
#> 95  095 -0.22939380 -1.227068215
#> 96  096  1.13137728 -0.177454673
#> 97  097  1.06855723  0.613845500
#> 98  098  0.93392884  0.904671393
#> 99  099 -0.10607048  1.344712655
#> 100 100 -1.63128117 -0.668811524

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.838825895 -0.59135323
#> 2   002 -1.549870197 -1.34687034
#> 3   003 -0.217062664  1.03628781
#> 4   004 -0.883281417 -0.37137660
#> 5   005 -1.208136338  0.19588830
#> 6   006  0.768093494  0.51782900
#> 7   007  0.012578772  0.95007192
#> 8   008 -0.653398426  0.06187765
#> 9   009  1.997569489  1.46009908
#> 10  010  1.180272909  1.02353380
#> 11  011  1.198900832  0.01636095
#> 12  012 -0.160976878  0.22101253
#> 13  013  1.696408968  0.97605804
#> 14  014 -1.101695722 -0.66889499
#> 15  015  0.819551548  0.10102012
#> 16  016  0.195434017  0.35311480
#> 17  017 -1.260333621  0.76848853
#> 18  018 -0.649355452 -0.61601258
#> 19  019  0.452061711 -0.41264505
#> 20  020 -1.749668380 -1.91496273
#> 21  021  0.991458459  1.53118394
#> 22  022 -0.007800676  0.46861693
#> 23  023  0.191639203 -0.37165992
#> 24  024  0.702601577 -0.06936482
#> 25  025  0.698027184 -0.27269912
#> 26  026 -0.786046637 -1.60177628
#> 27  027 -0.402730615 -0.67183275
#> 28  028 -0.543508677  0.77744010
#> 29  029  0.477142023  0.58168691
#> 30  030  1.248414717 -0.69037832
#> 31  031 -1.071519307 -0.11544555
#> 32  032  0.103669621  0.21019485
#> 33  033  1.049434254  2.02980512
#> 34  034 -0.447090427 -0.51039419
#> 35  035  1.078056647  0.63872519
#> 36  036  0.605649561 -0.96406493
#> 37  037  0.047375280  0.03432006
#> 38  038  0.584577147 -0.56972068
#> 39  039  0.703186058  0.44134336
#> 40  040  0.282702041  0.36862326
#> 41  041 -0.079448929  1.08724405
#> 42  042  1.389335425 -0.09640418
#> 43  043 -0.805151923  0.88869953
#> 44  044 -1.628543814 -0.55423374
#> 45  045  0.332922736  0.70592431
#> 46  046 -0.425977903  1.38966937
#> 47  047 -0.483169247 -0.29027193
#> 48  048 -0.559827076  0.82746011
#> 49  049 -1.168119616 -0.41974941
#> 50  050 -1.411934096  0.52307642
#> 51  051  1.347059078 -0.11934246
#> 52  052  1.255104919  0.18523777
#> 53  053 -0.147437810 -0.59683132
#> 54  054  0.499277504 -0.87363754
#> 55  055 -0.301852153 -0.15747120
#> 56  056 -0.870029723 -1.43625905
#> 57  057  0.758026311  0.61134091
#> 58  058  0.409773312  1.27909391
#> 59  059 -1.764978772 -0.63259833
#> 60  060 -0.340908795 -0.12304978
#> 61  061  0.210070826 -0.15897902
#> 62  062  0.180817886  0.28414934
#> 63  063 -0.722334046 -0.22829823
#> 64  064  0.754974991 -0.02018400
#> 65  065  0.388529102  1.66766164
#> 66  066 -0.588384783 -1.71567207
#> 67  067  0.638253150 -0.53261415
#> 68  068 -0.344916075  0.52085851
#> 69  069 -2.295508725 -1.55723684
#> 70  070  0.296949466 -0.19819587
#> 71  071 -1.335539870  0.15815880
#> 72  072  0.074039040 -0.08426890
#> 73  073 -0.270591013 -2.86034296
#> 74  074  0.149050187  0.56142478
#> 75  075  0.194776465  0.22886494
#> 76  076 -2.500829297 -1.03632202
#> 77  077 -0.870808360 -0.89729458
#> 78  078  0.648421107 -0.65442179
#> 79  079  0.018746554  0.53208127
#> 80  080 -1.427818428 -1.29098267
#> 81  081  1.971478528  1.56153813
#> 82  082 -1.387444935 -0.42713208
#> 83  083  0.090828356 -0.93798585
#> 84  084 -1.847189505 -0.38399816
#> 85  085  2.016798836 -0.58833015
#> 86  086 -0.855798912 -0.60175799
#> 87  087 -0.268929496 -0.79538898
#> 88  088 -1.175303137 -0.67131120
#> 89  089 -0.218039232 -1.72752300
#> 90  090 -1.381649545 -0.03806625
#> 91  091 -0.348910219 -0.36891664
#> 92  092  0.064417874 -1.62416146
#> 93  093 -0.160015484  0.20432799
#> 94  094  0.190612611  0.97044301
#> 95  095  0.228336380 -0.40653770
#> 96  096 -0.531911109  0.75871831
#> 97  097  0.132218693  2.40684345
#> 98  098 -0.173756259  0.85728112
#> 99  099  0.828032816  0.25112019
#> 100 100  0.781688282  0.81639156