Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4939401
#> Y_2 0.4939401 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.6448300 -0.6939003
#> 2 -1.1449384 -1.4339420
#> 3 -0.9644822 -1.0244086
#> 4 -0.7671550 -0.5846580
#> 5  0.1852119 -1.0130764

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.124023911  0.033480810
#> 2   002 -2.128533179 -1.390201694
#> 3   003  1.374284214  0.393902362
#> 4   004 -0.300301398  0.822030492
#> 5   005  1.428960462  1.387361666
#> 6   006 -0.050158562 -1.146739675
#> 7   007 -1.074721007 -1.222030967
#> 8   008 -1.367746032  0.736500066
#> 9   009  1.524271352 -0.388632230
#> 10  010 -0.412501831  0.371801488
#> 11  011 -0.668870848 -1.480271714
#> 12  012 -0.065250403  2.410314174
#> 13  013  1.184814142  0.681875265
#> 14  014 -1.160212997 -0.067323522
#> 15  015 -0.451777626 -1.530915259
#> 16  016 -0.888713965 -0.497535005
#> 17  017  0.455033506  0.152686812
#> 18  018 -0.752802408 -0.470878851
#> 19  019 -1.467305268 -1.206115630
#> 20  020 -0.121809256 -0.131495753
#> 21  021  0.780367086  1.583933906
#> 22  022 -3.120832332 -2.426378679
#> 23  023 -1.247089058 -1.050769933
#> 24  024 -2.795697523  0.196676613
#> 25  025  1.635550188  0.874781199
#> 26  026  1.143544337  0.451195297
#> 27  027  0.090174839 -0.275057759
#> 28  028 -1.131761783 -2.041091412
#> 29  029  2.103422291  2.994853411
#> 30  030  0.142437569 -0.788884321
#> 31  031 -0.448209452 -0.302126590
#> 32  032  0.185546845 -0.297388086
#> 33  033  0.031871499 -1.035834461
#> 34  034  1.323982144  1.254495269
#> 35  035 -0.844336476 -2.376212897
#> 36  036 -1.969826768 -0.860399024
#> 37  037  0.072420255  1.081727704
#> 38  038  0.066416691 -0.287219195
#> 39  039  1.073282427  0.489857218
#> 40  040  0.023912919  0.175394734
#> 41  041 -1.273848464 -1.271171690
#> 42  042 -0.539553748 -0.594775737
#> 43  043  1.265858884  0.481710933
#> 44  044  0.514570658  0.574174057
#> 45  045  0.559569766  0.865799986
#> 46  046 -1.211571003 -0.806101325
#> 47  047  0.838023658  0.436637447
#> 48  048 -1.648353616 -0.637422349
#> 49  049 -0.578676826 -0.327269886
#> 50  050 -0.994231487  0.083158829
#> 51  051 -2.102366930 -0.330274031
#> 52  052 -2.228811446 -1.429905229
#> 53  053 -0.497638489 -1.044447450
#> 54  054 -0.255462420 -1.353216809
#> 55  055 -0.231486628  0.344978515
#> 56  056  0.173847368 -0.510364269
#> 57  057 -1.830136913 -1.535936008
#> 58  058  1.086877870  1.402162089
#> 59  059  0.615515954 -0.888127525
#> 60  060  0.581416870  0.361010969
#> 61  061 -0.717806167  0.830645322
#> 62  062  0.318882631 -0.449876180
#> 63  063 -0.992984245  0.110687763
#> 64  064 -1.618535253  0.788628635
#> 65  065  1.143477246 -0.329157403
#> 66  066  0.563253390  1.589546022
#> 67  067  0.230843315  2.262576990
#> 68  068  0.444125760  0.998115801
#> 69  069 -0.117778409 -0.387828574
#> 70  070 -0.063561927 -0.088602208
#> 71  071  0.165696970  0.400829430
#> 72  072 -0.401982302 -0.201066698
#> 73  073  0.178806263  1.320311906
#> 74  074 -0.110840741 -0.507488619
#> 75  075 -0.110411242 -0.115195699
#> 76  076  0.002454124  0.410641104
#> 77  077  0.395059665  1.045120062
#> 78  078 -1.464182623 -0.972722412
#> 79  079 -0.469177823  1.916527108
#> 80  080  0.933734821  1.849758034
#> 81  081 -0.510490409 -0.933488109
#> 82  082 -1.635013526 -0.573264969
#> 83  083  0.042079242  1.604131444
#> 84  084  0.032161766  0.229968892
#> 85  085  0.316894779  0.662374455
#> 86  086  1.274543536  0.958800447
#> 87  087 -1.489241964 -1.365330466
#> 88  088 -0.368606233 -1.709202410
#> 89  089 -0.249740081  0.004871023
#> 90  090  0.488468785  0.779705599
#> 91  091 -0.756555984 -1.065095759
#> 92  092  2.089668759  1.634605876
#> 93  093  2.628878854  2.434039191
#> 94  094 -1.568701801 -0.744960183
#> 95  095  1.364629351  0.165268181
#> 96  096 -0.451123017  0.361323522
#> 97  097  0.517034187  0.209301825
#> 98  098 -1.681444204  0.340649514
#> 99  099 -1.153660815 -0.132582640
#> 100 100  0.446673818  0.426254138

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -0.514616756  0.363951267
#> 2   002 -0.574852460  0.838811908
#> 3   003 -0.373261060  0.741423220
#> 4   004  0.520381365  0.246727320
#> 5   005  1.135958200  0.155827235
#> 6   006 -1.025251172 -0.333916919
#> 7   007 -0.765496317 -0.309014151
#> 8   008  0.083914943 -0.289233626
#> 9   009 -1.160470036 -0.375644974
#> 10  010  0.643604760  1.119696160
#> 11  011 -0.546438297  0.551246324
#> 12  012 -0.700413858  0.081556529
#> 13  013  0.771651924  0.710085449
#> 14  014 -1.745169228 -0.828652321
#> 15  015 -0.593253622 -0.638342425
#> 16  016 -0.243572270 -0.396737181
#> 17  017  0.225946722  0.019590591
#> 18  018  2.008997480 -0.102040500
#> 19  019 -1.152077820 -0.139777436
#> 20  020  0.338386399 -1.118501095
#> 21  021 -0.718231306 -1.200817583
#> 22  022  0.438585737  0.341685586
#> 23  023  0.144967110 -1.588580475
#> 24  024 -0.569029884 -0.494675056
#> 25  025  1.191313890 -0.295811281
#> 26  026  1.522300686 -0.276490112
#> 27  027 -1.829196496 -1.542600554
#> 28  028 -1.154598576 -0.405930886
#> 29  029 -0.716320531 -0.225185508
#> 30  030 -0.446775130  0.452468899
#> 31  031 -0.064839870  1.282714806
#> 32  032  0.030239652 -0.891056686
#> 33  033  1.331947636  0.779770903
#> 34  034  2.164324787  1.894917537
#> 35  035 -0.776717486 -0.809011188
#> 36  036  1.042291176 -0.056859845
#> 37  037  0.606854731  1.022459945
#> 38  038  1.053895547  1.004833576
#> 39  039 -0.313723181  0.379525508
#> 40  040  0.644259432  0.727227976
#> 41  041  1.129469779 -0.279178215
#> 42  042 -1.059938768 -0.225399913
#> 43  043 -1.617250697 -1.245769781
#> 44  044  0.801724885 -0.661756219
#> 45  045 -0.860221254 -2.324479455
#> 46  046  0.528652314  0.484458899
#> 47  047  0.285235096 -0.343791977
#> 48  048 -0.950987442 -0.521120428
#> 49  049  0.192056412  1.443004184
#> 50  050  2.945931198 -0.089070809
#> 51  051  1.179460310 -0.502852824
#> 52  052 -0.007872431 -0.691554453
#> 53  053  0.548375142 -0.452850681
#> 54  054 -2.057702360 -0.882485563
#> 55  055  1.072434671  1.905573253
#> 56  056 -0.628314093 -0.373150980
#> 57  057 -0.704180583 -1.570496852
#> 58  058 -0.093205232  0.878048273
#> 59  059  0.373108560 -0.288449916
#> 60  060  0.609671767 -0.311193675
#> 61  061 -0.587593810 -0.603625252
#> 62  062  0.965919439  0.981154317
#> 63  063  1.373869988  0.459834219
#> 64  064 -1.423728976  1.462764021
#> 65  065  1.495953767  0.228763818
#> 66  066 -1.891040277 -1.278027053
#> 67  067 -1.401601816 -0.578059223
#> 68  068 -0.663197817 -1.692905377
#> 69  069  0.930664265  2.130552336
#> 70  070 -0.656378950 -0.533304306
#> 71  071 -2.527616072  0.544193834
#> 72  072 -0.232405237 -0.699602926
#> 73  073  1.306826531  2.565369958
#> 74  074  0.698487118  1.235865180
#> 75  075  0.093875906  0.034561451
#> 76  076 -1.010570414  0.009562386
#> 77  077  1.029551519 -0.026284792
#> 78  078  1.173208390 -1.782368363
#> 79  079  0.676271587  0.276369285
#> 80  080  2.127018760 -0.085407264
#> 81  081  0.029614140  1.291622830
#> 82  082 -0.042314330 -0.046330949
#> 83  083 -0.834525031 -0.476519422
#> 84  084 -0.317102673  0.838828215
#> 85  085 -0.011317520 -0.191110102
#> 86  086 -0.693207463 -0.946918458
#> 87  087 -0.338645562 -1.892581935
#> 88  088  1.810906712  0.827651479
#> 89  089 -0.382436131  0.798981612
#> 90  090 -0.387220189  0.364978631
#> 91  091  1.271080910  1.746758996
#> 92  092  2.496893424  0.369977618
#> 93  093  1.220618958  0.198657840
#> 94  094  0.480630531 -1.207197720
#> 95  095 -1.821942292 -2.757400669
#> 96  096  0.474923805  0.013538153
#> 97  097 -1.565326674 -1.086640854
#> 98  098  2.379948518 -0.381435396
#> 99  099 -0.880052920  0.794221258
#> 100 100  1.347169876  2.121007136