Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4783293
#> Y_2 0.4783293 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3602126  1.3101390
#> 2  0.2455751  0.5007854
#> 3 -2.2776163 -0.9753893
#> 4 -1.1528711 -1.9311550
#> 5  2.1589844  2.4667871

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.55436365 -0.67144253
#> 2   002  0.15477325 -0.58623533
#> 3   003 -0.94452512 -0.69989352
#> 4   004  0.98689347 -0.46061983
#> 5   005  1.76323807  1.03369297
#> 6   006 -0.26948074  0.26176550
#> 7   007 -0.32990615 -1.81645855
#> 8   008 -0.23476408 -0.35036609
#> 9   009  0.63455595  0.72090588
#> 10  010 -0.80658645 -1.44068326
#> 11  011 -0.84075906  1.29263665
#> 12  012  0.29268028  0.02961100
#> 13  013  0.77978567  0.16674500
#> 14  014  1.24857298  0.42527546
#> 15  015 -1.19224792 -0.89629303
#> 16  016  0.98563052  0.50103731
#> 17  017 -1.05105888  0.12537667
#> 18  018  0.48081611 -0.07060514
#> 19  019  0.42685931 -0.47058639
#> 20  020 -0.42737142  1.32923918
#> 21  021 -1.25768349 -1.32073246
#> 22  022 -0.94200755 -0.11015811
#> 23  023  1.98375453  1.66971215
#> 24  024  1.00085656 -1.11796921
#> 25  025  0.52459104  0.36550242
#> 26  026  0.49908383  0.07296654
#> 27  027  0.17290100  0.74642426
#> 28  028 -0.56079754  0.81138124
#> 29  029  0.88311461  0.17813782
#> 30  030  0.88012976  1.58523092
#> 31  031 -1.80687380 -1.58585075
#> 32  032 -1.05856584 -1.42233927
#> 33  033 -1.10780862  0.84071513
#> 34  034 -0.48996531 -0.77689728
#> 35  035  0.42128905  1.25672743
#> 36  036 -1.11784135  0.18428142
#> 37  037 -0.49408036 -0.04965095
#> 38  038  0.87944014 -0.06905661
#> 39  039 -0.69162954  0.95381074
#> 40  040 -0.92609883 -0.10065613
#> 41  041  0.58218716 -0.33904928
#> 42  042  2.19538098  1.34282038
#> 43  043  0.36765297 -0.62678205
#> 44  044  0.21889207  1.07293542
#> 45  045 -0.04073510  0.06706956
#> 46  046 -1.26977424 -0.11313742
#> 47  047 -0.62708601 -0.27099588
#> 48  048  0.41728341 -0.32595063
#> 49  049 -1.07363490  0.47951035
#> 50  050 -0.04210256  0.06067955
#> 51  051  0.91929576  0.39098118
#> 52  052  0.94140964 -0.37569062
#> 53  053 -0.25436987 -0.32440969
#> 54  054  0.73558941 -0.21001493
#> 55  055  0.91554765  1.61197495
#> 56  056 -1.14806558  0.01803070
#> 57  057  0.15030382  1.18176768
#> 58  058 -0.40179654  0.23907643
#> 59  059 -0.22759583  1.05422006
#> 60  060 -0.07612642 -0.29560029
#> 61  061 -0.46353631 -1.50470933
#> 62  062 -1.29656300  0.05779628
#> 63  063 -0.86268776 -1.15803538
#> 64  064 -0.36553724  1.02063595
#> 65  065 -1.39404806 -1.82810137
#> 66  066  1.18902325  0.96887900
#> 67  067  1.67617200 -0.90538533
#> 68  068 -0.76869630 -0.47012703
#> 69  069  0.45658754 -0.02702011
#> 70  070  0.14106927 -0.23659643
#> 71  071  1.53118780  1.30178442
#> 72  072  0.32444310  0.21376447
#> 73  073  1.10262215  0.31599249
#> 74  074  0.73645911 -0.48899821
#> 75  075  0.71069740 -0.27639744
#> 76  076 -0.99731241 -0.12004033
#> 77  077 -1.39728645 -0.10495051
#> 78  078  0.30465795  1.37162346
#> 79  079  0.87253863  0.10941989
#> 80  080  0.76288229  1.37039062
#> 81  081  2.23281775  1.60719531
#> 82  082 -0.94516149  0.02823757
#> 83  083 -1.12746905 -1.64691231
#> 84  084 -0.73503057 -0.80421025
#> 85  085 -0.14913235  0.18012166
#> 86  086 -0.49249012 -0.40219764
#> 87  087 -1.39747163  0.25093132
#> 88  088  0.02191799  0.50215079
#> 89  089 -0.03002510  0.10158652
#> 90  090 -1.28845665 -0.09163280
#> 91  091 -0.33080032 -1.04750107
#> 92  092 -1.95984153 -1.12032656
#> 93  093  2.22371526  1.19899900
#> 94  094  0.13783465 -0.88790905
#> 95  095 -0.50385168 -0.59803674
#> 96  096 -0.13782594  0.13835959
#> 97  097  0.54079342  0.86624298
#> 98  098  2.61527778  1.49251026
#> 99  099  0.17801023 -0.73484477
#> 100 100  0.66629455  0.81300208

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1           Y.2
#> 1   001 -1.120309166 -0.0977123309
#> 2   002 -2.430965597 -1.6525492404
#> 3   003 -0.551143513 -1.6708841835
#> 4   004  0.551626573  1.0732165080
#> 5   005 -0.641793281  0.2107862011
#> 6   006  0.319737545 -0.4073170998
#> 7   007 -1.395876945 -0.7363252822
#> 8   008  0.674765340  0.5224332707
#> 9   009 -1.486342566  0.8179050877
#> 10  010 -0.992131843 -0.8277560726
#> 11  011 -2.389653463 -0.9698278625
#> 12  012 -0.711790187  0.1411011122
#> 13  013  0.205129970  0.6277858270
#> 14  014 -0.411296239  0.3590932144
#> 15  015 -1.540459102 -0.9165329140
#> 16  016 -0.010101971  1.2082747065
#> 17  017  0.375126219  0.8562893585
#> 18  018  0.626119683  0.9571963073
#> 19  019  1.214196682 -0.7715030270
#> 20  020 -0.603482185 -0.9405823697
#> 21  021 -0.003703134  1.1640523241
#> 22  022 -0.416408130 -0.9528904450
#> 23  023 -2.330798393 -0.4060744479
#> 24  024  0.044662894 -0.0471497946
#> 25  025 -0.271141691 -0.1988037515
#> 26  026  0.590739573  0.3737129730
#> 27  027 -1.625379401 -1.8189803739
#> 28  028  0.581422552  0.3265410470
#> 29  029  0.720511726 -0.9799588102
#> 30  030  0.131900915  0.1880969340
#> 31  031  1.275127724  1.1667380804
#> 32  032 -1.009087997 -0.0943194762
#> 33  033  1.228231140  2.2074989536
#> 34  034 -0.510805062 -0.0122034887
#> 35  035  0.604384727  0.1569294649
#> 36  036  0.754807867 -1.4830771195
#> 37  037  2.240055872  0.8589392103
#> 38  038  0.059736484  1.4067341682
#> 39  039 -0.581100397 -0.1005931841
#> 40  040 -0.272090846 -0.2165726752
#> 41  041 -0.432468866 -0.2293331769
#> 42  042  0.624347128 -0.5268927807
#> 43  043 -0.271378037 -0.6018435774
#> 44  044  0.608160657  0.2521090368
#> 45  045 -1.718650356 -0.7596088770
#> 46  046  2.374351394  1.7520814024
#> 47  047 -0.102966222  0.9373223822
#> 48  048  1.371453367 -1.0520266308
#> 49  049  1.280114477 -0.1810537888
#> 50  050 -1.297191909 -1.2822625989
#> 51  051 -0.861426080 -0.5460592829
#> 52  052  0.974149317 -0.3924523965
#> 53  053  0.055958582 -0.0007911308
#> 54  054 -1.306014230 -0.5272047189
#> 55  055  1.235289852  0.1632169911
#> 56  056 -1.298459435 -1.3495675181
#> 57  057  0.498860234 -0.1347891207
#> 58  058  1.489532137  0.4286795442
#> 59  059 -0.772697414 -0.5728848558
#> 60  060 -0.302123467  0.8772209662
#> 61  061 -1.245201261  1.0639879859
#> 62  062 -0.086966067  0.0763423485
#> 63  063 -1.255074493  0.4088349618
#> 64  064 -0.178185214 -1.0376928663
#> 65  065  1.845713287  0.1492643049
#> 66  066  1.813717673 -0.1091911470
#> 67  067 -1.901479006  0.5423701018
#> 68  068  0.644729982  1.3635085771
#> 69  069 -0.657788340  1.2722036546
#> 70  070 -0.853757165  0.6534334746
#> 71  071  0.046359576 -1.0978237966
#> 72  072  1.262852043  1.5456966505
#> 73  073 -1.373332682 -0.6360860765
#> 74  074 -0.017999732  1.4542315217
#> 75  075 -0.233645423 -0.9027563524
#> 76  076 -1.618023415 -0.4854063073
#> 77  077  0.074505997  0.0677377051
#> 78  078  0.410966557  0.3704063624
#> 79  079 -1.351799124  0.3458365908
#> 80  080 -1.175152951  0.1196266357
#> 81  081  0.029605960 -0.4530989880
#> 82  082 -0.052360121 -1.3289234444
#> 83  083  1.944074932 -0.0646083933
#> 84  084 -1.394357943 -1.2371722075
#> 85  085  2.016424707  1.0317284724
#> 86  086 -1.229498890 -1.7552275487
#> 87  087  2.494182155  1.7182807688
#> 88  088  1.085581316  0.7951673860
#> 89  089  0.100163986  0.3658723332
#> 90  090  2.302909974 -0.2252077571
#> 91  091  1.332443168  0.3156830515
#> 92  092  1.102509782  0.7661180146
#> 93  093 -1.077643943 -0.4179271978
#> 94  094 -0.908629946  0.3659452734
#> 95  095  0.716430185  0.4408734572
#> 96  096  0.042350326  0.7338794815
#> 97  097  0.452489138 -0.8785520324
#> 98  098  1.796390378  1.0824427502
#> 99  099  0.975527939 -0.0923379946
#> 100 100 -0.950954524 -1.0967111315