Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5213123
#> Y_2 0.5213123 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3003470 -0.5694663
#> 2 -1.0400107 -0.4830231
#> 3  0.7765792  1.5585074
#> 4  0.2831390 -0.5202510
#> 5  0.6486046 -0.7026492

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  2.21924158  1.930923968
#> 2   002  0.24180811 -0.236110789
#> 3   003  0.21582670  1.330074176
#> 4   004  0.40150454  0.512038107
#> 5   005 -1.09640442 -2.182844230
#> 6   006  0.65023100 -0.444544465
#> 7   007  1.49398082  0.349418851
#> 8   008 -1.07706149 -0.714983397
#> 9   009  0.07684348  0.879356084
#> 10  010  1.33293856  2.710091153
#> 11  011 -1.14642772 -1.560618280
#> 12  012 -0.23693815  0.560116860
#> 13  013 -0.57221300 -1.574855271
#> 14  014  0.27223742 -0.098342420
#> 15  015  1.44056938  1.429643341
#> 16  016  0.50767369 -0.351263998
#> 17  017  0.41518656  0.032066392
#> 18  018  1.24443578 -0.397050733
#> 19  019  0.51465733  0.476668104
#> 20  020 -1.62361411 -1.470004151
#> 21  021  0.48934111  0.185146320
#> 22  022 -0.62942898  0.527249853
#> 23  023 -0.34154983  0.701965463
#> 24  024  0.61207572 -0.928702953
#> 25  025 -0.58076150 -0.487542770
#> 26  026  1.19490437  0.269380579
#> 27  027  0.79047835  0.703514797
#> 28  028  0.30973861 -0.787535091
#> 29  029  0.06479496  0.746326164
#> 30  030 -0.18896992 -0.773497260
#> 31  031 -0.15731601  0.438968790
#> 32  032 -0.69781187  0.182644159
#> 33  033 -0.36352048 -0.776136376
#> 34  034  0.11540877  0.501311900
#> 35  035  0.79246742  0.286082153
#> 36  036  0.31893490  1.420555338
#> 37  037 -1.00548148 -0.977698547
#> 38  038 -0.40491096  1.598475527
#> 39  039  1.27445435 -0.773728685
#> 40  040  0.71789358  1.000412788
#> 41  041  0.50996054 -0.427928404
#> 42  042  0.34871680  0.938474802
#> 43  043  1.14450291  1.898354706
#> 44  044  0.62331284 -0.663017464
#> 45  045 -1.93551037 -1.471449954
#> 46  046  0.28076340 -0.264913225
#> 47  047 -0.39922821 -0.150356304
#> 48  048  0.55151216  0.694408053
#> 49  049 -0.03584087 -0.141479899
#> 50  050  0.53890216 -0.419197679
#> 51  051  0.67807581 -0.363232902
#> 52  052  0.58724750  1.267396014
#> 53  053  1.80208027  0.496036944
#> 54  054  1.02187723 -1.236276831
#> 55  055  1.61502162  1.345312784
#> 56  056 -1.54480962 -1.464282792
#> 57  057 -0.14250354 -0.555676277
#> 58  058  1.02417535  0.977689089
#> 59  059  0.11977334 -1.022348286
#> 60  060  1.51760731  0.315008449
#> 61  061 -1.43742497 -0.493735010
#> 62  062 -0.78007236 -0.548704813
#> 63  063  0.18706092  0.354745305
#> 64  064 -1.30907471 -0.839513855
#> 65  065  0.81509647  0.353644726
#> 66  066 -0.27548761  1.858521429
#> 67  067 -0.01741959 -0.328354367
#> 68  068  0.91760749  1.562020521
#> 69  069  1.21415205 -0.380039591
#> 70  070 -0.73783105 -2.304136026
#> 71  071 -0.26295756  0.344996951
#> 72  072  0.65064666  1.597408796
#> 73  073 -0.49787851  0.567552460
#> 74  074  0.37261106  0.103359653
#> 75  075  0.25757289  0.805708395
#> 76  076 -0.09298703 -0.122621654
#> 77  077  0.68359334  1.816888498
#> 78  078 -0.49753474 -0.584895644
#> 79  079  0.39102938  0.090095756
#> 80  080 -0.63103168  0.593472310
#> 81  081 -1.44091058 -0.316256713
#> 82  082 -0.80865246  0.007069463
#> 83  083 -0.66045239  0.893831233
#> 84  084  1.98620083  1.553785795
#> 85  085  1.77086335  1.745747064
#> 86  086 -0.05306469 -0.772821487
#> 87  087  1.24918085 -0.794688435
#> 88  088 -0.34230118 -2.078888723
#> 89  089  0.13586302 -1.128171617
#> 90  090  0.61178571 -0.111158208
#> 91  091 -0.48068540 -0.574894643
#> 92  092  1.36321802  0.662314839
#> 93  093  1.47757095  2.862018735
#> 94  094 -0.18160611 -0.063443461
#> 95  095  1.70717768  0.014695435
#> 96  096 -0.14777981 -0.231016410
#> 97  097 -1.24610519  0.383738568
#> 98  098 -0.51946697 -0.895420743
#> 99  099  1.01790986 -1.120539560
#> 100 100 -0.04262985  0.104130531

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  0.5023495735  1.15234404
#> 2   002 -0.1278699819 -1.32301691
#> 3   003 -0.6502162889 -1.02993650
#> 4   004  1.0987586013  0.08625937
#> 5   005  0.1077113732 -1.01172717
#> 6   006  1.1674031045 -0.15306139
#> 7   007  0.1338036066  0.82543626
#> 8   008  1.4095443380  1.55355947
#> 9   009  0.7907381437  0.15672996
#> 10  010 -0.5550508206  0.51240524
#> 11  011 -0.3235633839 -1.69117064
#> 12  012 -0.5484231030 -0.77395785
#> 13  013  0.3386435245 -0.54138880
#> 14  014  0.4946430227  0.76106482
#> 15  015 -1.5217588004 -0.48015365
#> 16  016 -1.2889993733  0.18071113
#> 17  017 -0.1276274962  0.40590253
#> 18  018  0.4097220629 -0.09178657
#> 19  019 -0.3299498176 -0.44511410
#> 20  020 -0.3460384830 -0.97714479
#> 21  021 -0.1953483629 -1.69606810
#> 22  022 -1.4933276300 -0.89962858
#> 23  023  0.5429891454 -1.33416869
#> 24  024  0.1621758556  0.89168538
#> 25  025 -1.9615979281 -0.70709429
#> 26  026 -1.0307945590 -0.43820975
#> 27  027 -0.8032764160 -1.00488629
#> 28  028 -0.4870869587 -0.39318948
#> 29  029  0.9724332262  2.22757333
#> 30  030 -0.9724992009 -0.88809330
#> 31  031  0.2193743499 -0.50848817
#> 32  032 -0.5744222775 -0.83663174
#> 33  033  0.1662832760 -0.06613579
#> 34  034 -0.1790346015 -0.86218204
#> 35  035 -1.4780996436 -1.66835586
#> 36  036 -2.1402147539 -1.48498553
#> 37  037 -0.4725826591  0.71897424
#> 38  038  0.8135016034  0.46689927
#> 39  039 -1.4187036452 -1.30286473
#> 40  040  1.1537076588 -0.92854219
#> 41  041 -1.2007747318 -0.97337353
#> 42  042 -0.1675137654 -1.13758002
#> 43  043  0.4094926136 -0.20169072
#> 44  044  1.0860834168  1.00484612
#> 45  045  0.7317158940 -2.08245294
#> 46  046  2.7691408172  1.78014349
#> 47  047  0.3353329188 -0.19699131
#> 48  048  1.9605452868  1.96374627
#> 49  049  0.9139345676  0.03883440
#> 50  050  0.0373844655  0.37406944
#> 51  051 -1.7298814822 -2.99358112
#> 52  052 -0.0253550621 -1.23006623
#> 53  053 -0.0847722856 -0.79335443
#> 54  054 -0.2271543741 -0.79762087
#> 55  055 -1.8293355135  0.57359840
#> 56  056 -0.0090299812  0.98953308
#> 57  057 -0.1672438599 -0.71958078
#> 58  058  0.7131894633 -1.07255149
#> 59  059  1.1137366760  1.38291088
#> 60  060 -0.9518685920 -0.03321179
#> 61  061  2.5198386620  0.61366080
#> 62  062  0.1884296654  1.32081265
#> 63  063  1.3904191694  1.65217245
#> 64  064  1.2190772844  0.69980952
#> 65  065  1.2615980102  0.71876875
#> 66  066 -1.2235130399 -0.99454648
#> 67  067 -0.0835463655  1.72678374
#> 68  068 -0.0030553673 -0.08779144
#> 69  069  0.2094274277  0.12624528
#> 70  070  0.0286116859 -0.34876368
#> 71  071 -1.2597649121 -2.26659723
#> 72  072  1.2489455965 -0.67338913
#> 73  073  0.5132753418 -1.70943242
#> 74  074 -0.3717314203 -0.26428798
#> 75  075 -0.5242563045 -1.28717107
#> 76  076 -1.9068661408 -3.00662859
#> 77  077 -0.3659081153  0.25250751
#> 78  078 -0.2459923529  2.09307556
#> 79  079  0.1494648289 -0.10125206
#> 80  080  0.8667621345  0.92957556
#> 81  081 -0.5042029767 -1.04336711
#> 82  082  1.2554018005 -0.06061036
#> 83  083  0.0596755585  1.02366436
#> 84  084 -2.6329474724 -1.51192785
#> 85  085 -0.0575940642 -0.01306034
#> 86  086 -0.5885584609  0.34709316
#> 87  087  1.1965796743  1.91384865
#> 88  088 -2.4336933306 -2.65623127
#> 89  089  0.2517996329  1.28386649
#> 90  090  1.1581828225 -0.08972740
#> 91  091  0.1672716778  0.04022900
#> 92  092 -0.0003253716  0.98900308
#> 93  093 -0.2162132815 -0.04226425
#> 94  094 -0.1741921869  0.33123951
#> 95  095  0.3830343903  1.52687465
#> 96  096 -0.8957616000 -0.37156840
#> 97  097 -0.7214693559 -0.57313145
#> 98  098  1.2557140218 -1.08777564
#> 99  099 -0.1472757264  0.03437357
#> 100 100 -0.6787140337 -0.57851318