Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5691737
#> Y_2 0.5691737 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -2.4543965 -1.0725788
#> 2 -0.7568321  0.0613672
#> 3  0.2965908  1.1522955
#> 4  2.0295043  2.3380861
#> 5  0.5386867 -1.0609623

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.31516933  0.71174657
#> 2   002  0.05566736  0.46472135
#> 3   003  0.74208981  1.00567286
#> 4   004  0.30364404 -0.72356982
#> 5   005 -1.30368405 -2.07765747
#> 6   006  0.12702006  0.09448131
#> 7   007 -0.68123725 -1.11070963
#> 8   008 -1.11594752 -0.01931999
#> 9   009  1.48275696  1.11876854
#> 10  010 -1.32858153 -1.58217084
#> 11  011  0.72371401  0.13991915
#> 12  012 -1.15517846 -0.75865242
#> 13  013  0.94491181  0.18257047
#> 14  014 -1.93451165 -1.93713051
#> 15  015 -1.27679695  0.29591148
#> 16  016 -1.24920472  0.32959555
#> 17  017  1.32099086  1.79782392
#> 18  018 -1.06349817 -0.05183322
#> 19  019 -0.79636223 -1.38472265
#> 20  020  0.22895906 -1.15655840
#> 21  021  0.47654374  0.40752288
#> 22  022  0.56395951  0.64459083
#> 23  023 -0.06374620  0.65530529
#> 24  024 -0.20808235 -0.44454742
#> 25  025  0.03092286 -0.77582222
#> 26  026  0.14251582  0.61596472
#> 27  027  0.85707397 -1.46945287
#> 28  028 -0.79069338 -0.30998695
#> 29  029 -0.34883806  0.38793927
#> 30  030 -0.10303827  0.44148573
#> 31  031 -0.70522937 -0.18328293
#> 32  032 -0.15238172 -0.46078100
#> 33  033 -0.52943030 -0.63602852
#> 34  034  1.77526788  0.18828436
#> 35  035 -0.74962351 -1.68787642
#> 36  036  1.56291799  0.64921972
#> 37  037  0.07391753 -0.89827126
#> 38  038  0.40834896 -0.11428960
#> 39  039  0.36114567  1.88021869
#> 40  040 -0.81628473 -0.49064502
#> 41  041 -0.91787223 -0.88000020
#> 42  042 -0.33253947 -0.45144395
#> 43  043 -1.64867826 -1.83625763
#> 44  044  0.59511443  0.06947017
#> 45  045  1.10856289  0.87391163
#> 46  046 -0.75956413  1.10769084
#> 47  047 -0.78772848 -0.78631726
#> 48  048  0.55350478  0.22267506
#> 49  049  0.67571452  1.62260274
#> 50  050  1.42258782 -0.04720878
#> 51  051  0.61359915  0.41841804
#> 52  052 -0.32228371 -0.76802794
#> 53  053 -0.34121866 -0.50932659
#> 54  054  1.89091953  1.83213013
#> 55  055 -1.94140706 -0.25323109
#> 56  056 -0.07491878 -0.66560905
#> 57  057 -0.95648129 -0.61777849
#> 58  058  1.48570343 -0.75643260
#> 59  059 -1.51276363  0.24091870
#> 60  060 -0.48683837 -0.15807657
#> 61  061 -1.07503665 -0.51594176
#> 62  062  0.80919106  0.59180383
#> 63  063  1.58427191  0.72066094
#> 64  064  0.44707032 -0.25419957
#> 65  065  0.54079060  1.16845709
#> 66  066  0.52974476 -0.88488430
#> 67  067 -0.75240908  0.58580933
#> 68  068 -0.31695585  0.32197659
#> 69  069 -0.15076313 -0.57887235
#> 70  070 -0.17402648 -0.34301715
#> 71  071  0.59497798 -0.11649810
#> 72  072 -0.59326788 -0.36026683
#> 73  073  0.20303168  0.45604179
#> 74  074 -2.52741436 -2.62394604
#> 75  075 -1.47934312 -1.13527166
#> 76  076  0.45279422 -1.04505893
#> 77  077  0.84447186 -0.84912390
#> 78  078  0.58011750 -0.68275425
#> 79  079  1.11421056 -0.38569889
#> 80  080 -1.17522823 -0.61200533
#> 81  081 -0.21340169 -0.28254943
#> 82  082  0.59793818 -0.78917031
#> 83  083  0.77289636  0.52786641
#> 84  084 -0.06547352  0.72704521
#> 85  085  0.25311676  0.43151004
#> 86  086  1.17930510 -0.13218734
#> 87  087  1.22189058  1.93578663
#> 88  088 -0.52057059 -1.21133922
#> 89  089 -0.74662931 -1.64375003
#> 90  090 -0.13857419 -0.56163498
#> 91  091 -1.31958266 -1.06077249
#> 92  092 -0.51496964 -0.17477784
#> 93  093  0.14017519 -0.30240425
#> 94  094  0.81458169  0.07744391
#> 95  095  0.90545601  1.06501279
#> 96  096 -0.32973861 -1.10481880
#> 97  097  0.67463470  0.24198071
#> 98  098 -0.40112216  0.04487278
#> 99  099 -0.22445091 -1.71273132
#> 100 100 -0.51086852 -1.10564666

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1          Y.2
#> 1   001  0.8241061139  1.316609761
#> 2   002  0.9460165524  1.243066528
#> 3   003  0.8111603184  0.116718995
#> 4   004  0.0560125161  1.074647161
#> 5   005  0.3771547943 -0.354589848
#> 6   006 -0.9112958340 -0.345130685
#> 7   007 -2.1521446353 -1.502584151
#> 8   008  0.7425142996 -0.241722196
#> 9   009  0.4447453294  0.713166114
#> 10  010  0.0153240825 -0.070427217
#> 11  011 -0.5413828083 -1.905351213
#> 12  012  0.7662948642  1.009738486
#> 13  013  0.5578315680 -0.679985365
#> 14  014  0.9350422404  1.542508385
#> 15  015 -1.4986376484  0.016741120
#> 16  016 -0.8021377529 -0.654675679
#> 17  017 -0.5362618558  0.559007743
#> 18  018  0.6671425501 -0.921378268
#> 19  019  0.4508809291 -0.905434759
#> 20  020 -0.4224203222 -0.420584372
#> 21  021  1.0238192585  1.648802395
#> 22  022  0.0965728783  0.637187305
#> 23  023 -0.3310484602  0.051054307
#> 24  024  0.0447656595 -0.474216002
#> 25  025  0.1564831588 -0.218874993
#> 26  026  0.1450335238  0.428797342
#> 27  027  1.4979915951  0.285138613
#> 28  028  0.6213498391 -0.110160192
#> 29  029  0.2287363920  0.430403376
#> 30  030 -1.5158503390 -2.159657845
#> 31  031  0.0022115251  0.437960129
#> 32  032  0.3218108784  0.240517078
#> 33  033  0.1197241622  0.085170123
#> 34  034  0.9802910587  0.757721965
#> 35  035 -0.2280399267 -0.278674488
#> 36  036  0.5454643456 -0.046694151
#> 37  037  1.0140749763  1.345122687
#> 38  038  0.5292937825 -0.359277061
#> 39  039 -0.8335803973  0.004600125
#> 40  040  0.7640358215  0.712000506
#> 41  041  0.4176063787 -1.281182762
#> 42  042  0.6385715348 -0.516109218
#> 43  043  0.4699834317 -1.770186079
#> 44  044 -0.6814850773  0.665797436
#> 45  045 -0.7165899447  0.406775675
#> 46  046 -0.7968400952 -0.399540899
#> 47  047 -0.2555205456  0.706988183
#> 48  048 -0.4487654016 -0.273496677
#> 49  049  0.0005871078  0.955442871
#> 50  050  0.5204143379  0.586499083
#> 51  051  0.7889821092  0.954999817
#> 52  052  1.0271207806  0.199003193
#> 53  053  1.1642634897  0.426014643
#> 54  054  0.0088414966 -1.075327049
#> 55  055 -0.2128429849 -0.488888378
#> 56  056 -0.9854017747  1.231028769
#> 57  057 -0.6392251730 -1.416073343
#> 58  058  1.0716447220  1.812726436
#> 59  059  3.2175241593  1.062434098
#> 60  060 -0.3217965597 -2.374161129
#> 61  061 -0.1145991333 -0.850821934
#> 62  062  0.9818433387  1.741431590
#> 63  063 -2.1869937649 -1.058196074
#> 64  064  0.5817140880  0.803627236
#> 65  065 -1.3443649504 -0.827363336
#> 66  066 -0.2482065981  0.612188920
#> 67  067 -0.4591756936 -0.166839696
#> 68  068 -0.7901763581 -0.153875124
#> 69  069 -0.8860354899 -0.085871622
#> 70  070  2.1852257015  1.345054060
#> 71  071  1.3973235720 -0.544308129
#> 72  072  0.4527993625  0.608676666
#> 73  073 -1.7169162752 -1.210808666
#> 74  074 -0.8508916443 -0.411950102
#> 75  075 -0.6913564381  0.160378570
#> 76  076 -0.6398827865 -0.759088233
#> 77  077  0.5491655297 -0.144637821
#> 78  078  1.0825643506 -0.359279026
#> 79  079 -0.9082154085 -1.186501169
#> 80  080 -0.5513385373 -1.237251825
#> 81  081 -0.8208425896  0.512106576
#> 82  082  1.6071250142  1.287395638
#> 83  083 -0.5561649306  0.511008619
#> 84  084  0.1571945452 -0.349786603
#> 85  085  1.1404599702 -0.400383143
#> 86  086 -0.9580802675  0.996280620
#> 87  087 -0.7126969212 -0.508068629
#> 88  088  0.0056048098  1.215283230
#> 89  089 -0.4099544140 -0.065674270
#> 90  090 -0.6005602648 -1.507260016
#> 91  091 -0.7021965435 -0.290573766
#> 92  092 -0.0416566346  0.823945981
#> 93  093 -0.0395416562 -1.638426724
#> 94  094  1.1199691734  0.894105020
#> 95  095  0.9056840815 -0.055188726
#> 96  096 -0.2558925939 -1.605351178
#> 97  097 -0.5369938773  1.149867010
#> 98  098  1.6153593199  1.909996332
#> 99  099 -0.3460184961 -0.142470473
#> 100 100  0.7166961105  1.594153752