Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4941505
#> Y_2 0.4941505 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.2496881 -0.8920977
#> 2 -1.1297564 -1.5986064
#> 3 -0.3680809  1.0522158
#> 4  0.9838295  0.8661987
#> 5 -0.9671504 -0.5463193

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1         Y_2
#> 1   001 -1.1730606741 -0.23932698
#> 2   002 -0.6833925736 -1.64022426
#> 3   003  2.1779204166  2.09398867
#> 4   004  0.1199103898 -0.40533089
#> 5   005 -0.8588194168 -0.10199873
#> 6   006 -0.1535474978 -0.14001314
#> 7   007 -0.0007684811 -0.82702938
#> 8   008  0.4235340567 -0.06217514
#> 9   009 -1.4021929118 -1.80103093
#> 10  010  0.3519093062  0.95066215
#> 11  011  0.1082593820  0.35296203
#> 12  012  0.1133548254 -1.19183725
#> 13  013 -2.7723800754 -3.17147892
#> 14  014 -0.4007541369 -0.77952123
#> 15  015  0.2699038574  0.42231188
#> 16  016  0.4922129611  1.04405050
#> 17  017 -0.2521974127 -1.18354007
#> 18  018  2.2851625680  0.74828597
#> 19  019 -0.5116744511 -0.06421230
#> 20  020  1.1163678196  1.23891833
#> 21  021  1.0635438027 -0.36763691
#> 22  022 -2.0752341957 -1.85347170
#> 23  023 -1.3584236483 -1.66221280
#> 24  024 -0.6833195384 -0.80105707
#> 25  025  0.7375688791 -0.61336268
#> 26  026 -0.4401074740 -1.97123232
#> 27  027  0.4272080686  0.33054765
#> 28  028  0.3519312546 -1.28940085
#> 29  029 -0.5671239408 -0.14176375
#> 30  030  0.6687292088 -0.29662283
#> 31  031 -0.3943003535  1.00422908
#> 32  032 -1.2199839058 -1.90685526
#> 33  033 -0.3805939178  0.34779549
#> 34  034 -0.6716204767 -2.12928651
#> 35  035 -0.6642054152 -0.59726232
#> 36  036  0.3768265839  0.56324989
#> 37  037  0.0058982310  0.03501598
#> 38  038  0.0241738267  1.11975764
#> 39  039  0.5178169917 -0.03845451
#> 40  040  0.0658830951 -0.88668618
#> 41  041 -1.3449239299 -1.21877807
#> 42  042  0.8284037820 -0.15636309
#> 43  043  2.4094390046  0.53689548
#> 44  044 -0.1676743085 -0.91212576
#> 45  045  0.1044312928  0.03034122
#> 46  046  0.3946726855  0.77553267
#> 47  047  1.7438590287  1.21338778
#> 48  048 -0.6117544603 -0.18304734
#> 49  049  0.9667502403 -1.07991428
#> 50  050 -0.2465648269  0.39677874
#> 51  051 -1.3450168036  0.16887416
#> 52  052 -0.0626982188  0.12163389
#> 53  053  1.0886391536 -0.63438698
#> 54  054  1.8053788050 -0.07513695
#> 55  055 -1.3945309933 -0.42551691
#> 56  056 -0.3466702746 -0.99020620
#> 57  057  1.5064559723  2.04652620
#> 58  058 -0.0835157362  0.12715807
#> 59  059 -1.2086327593  1.00400457
#> 60  060 -0.4810336358  0.24802847
#> 61  061 -0.0773283597 -0.26016387
#> 62  062  0.1748609740 -0.49690562
#> 63  063  0.5611339185 -0.51990794
#> 64  064  0.9068102482 -0.90846304
#> 65  065 -0.0989663288 -0.36884376
#> 66  066 -0.4094120994  0.03326141
#> 67  067  1.9718175996  0.97347575
#> 68  068 -0.3868069386  0.93002026
#> 69  069  1.1379636834  0.89300956
#> 70  070 -0.6606652566  1.03171604
#> 71  071 -0.3169963990 -0.52619792
#> 72  072 -0.0482990022  0.65152499
#> 73  073  0.8731980412 -0.77125620
#> 74  074  1.2426640691  2.32510047
#> 75  075 -0.9554627247 -1.18128038
#> 76  076 -1.6640880797 -2.16670436
#> 77  077  0.0169105704 -0.21365353
#> 78  078 -0.1781656478  0.10990099
#> 79  079  0.2131814440 -0.87688853
#> 80  080 -0.2287382038  0.84016103
#> 81  081 -0.1058482337 -0.59331328
#> 82  082  1.7452292817  0.33651937
#> 83  083  0.2861971117  0.39266759
#> 84  084  0.3748604467 -1.72727181
#> 85  085 -0.9523195669 -0.02634120
#> 86  086  0.9695620677  0.44388904
#> 87  087  1.1330535361  1.12287680
#> 88  088  0.6054954694 -0.40937128
#> 89  089  0.8125961778 -1.24798766
#> 90  090 -0.1854452851 -0.79765745
#> 91  091 -1.0342867466 -1.68236992
#> 92  092  1.4247875100  1.53482894
#> 93  093  0.4020187265  0.82396386
#> 94  094 -1.8480220392 -1.27541580
#> 95  095  2.2322878961  1.31692383
#> 96  096  1.6918037756 -0.11302944
#> 97  097 -0.1560944454  0.07911565
#> 98  098  0.4333125251  0.23681730
#> 99  099 -1.9340499015  0.03023938
#> 100 100  0.5567824308 -1.03638513

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1           Y.2
#> 1   001  0.5693858189 -0.8745074355
#> 2   002 -0.4367917389 -0.1465005805
#> 3   003 -1.8680626637 -0.5463124465
#> 4   004  1.7004788023  0.0001621331
#> 5   005  0.3670098421 -0.3357307213
#> 6   006  1.1046523034 -0.7632173379
#> 7   007 -0.4165505151  1.1499280221
#> 8   008 -0.9599618614 -1.0288294110
#> 9   009 -1.0932650140  0.1901403813
#> 10  010  1.2364193778  1.4696698689
#> 11  011  2.3483179214  2.1648010915
#> 12  012 -1.9623700766 -1.9680114319
#> 13  013 -1.0769398254 -1.3715213109
#> 14  014  0.1782052473 -0.7668673991
#> 15  015 -0.3021937430 -0.4957306515
#> 16  016  0.3394727699 -0.2769136849
#> 17  017  0.0754911548  0.5459002915
#> 18  018  1.0700915671  1.4115955485
#> 19  019 -0.1956935529 -0.7784924415
#> 20  020 -0.8155448920 -0.9916927323
#> 21  021 -0.5120613316 -0.0466105889
#> 22  022  0.4756629156  1.2896916623
#> 23  023  0.4785354878  0.3820534399
#> 24  024 -2.0169576259 -0.3454811848
#> 25  025  0.0411168125 -0.1400267277
#> 26  026 -0.4299416761 -1.1975112298
#> 27  027  1.4029407746 -0.2840620219
#> 28  028 -0.0513283710  0.9603888741
#> 29  029  0.8403133896  0.5251061776
#> 30  030 -0.8437468986 -0.2704266015
#> 31  031  0.2378050336  1.5250319563
#> 32  032 -0.7096112116 -1.1367320097
#> 33  033  0.3927351746  0.0058246632
#> 34  034  0.0797016990  0.5057587920
#> 35  035  1.4997974199  0.3265348799
#> 36  036 -1.8974804718 -1.6838241841
#> 37  037 -1.2949533778 -1.9031713002
#> 38  038 -1.3037160031 -2.0200735842
#> 39  039  0.5890167701 -0.4012417478
#> 40  040  0.2801855883  1.4077788083
#> 41  041  0.0067143866 -0.0881394946
#> 42  042 -1.1675158621 -1.4633707037
#> 43  043 -1.6795011703 -1.3135600711
#> 44  044  0.7101094068  1.4293097636
#> 45  045 -0.9142393235  0.7083468012
#> 46  046  0.5894076160 -1.8318770441
#> 47  047  0.1609748082  0.6992375746
#> 48  048 -0.1093665148 -0.2149818520
#> 49  049 -1.7603020993 -0.8332495490
#> 50  050 -1.3399310282 -1.6393660624
#> 51  051 -0.6313147690 -1.2110197957
#> 52  052  1.3112412903 -0.0562026442
#> 53  053 -0.6870426403 -0.1794709795
#> 54  054 -0.3882579592  0.0264246868
#> 55  055  0.1166362347 -0.5068129101
#> 56  056  0.6996207199 -0.5416838293
#> 57  057  1.3880378303  0.4871276961
#> 58  058 -0.1023896418  0.8757801191
#> 59  059  0.8288776387 -1.6766316926
#> 60  060  0.1443915735  0.1571211098
#> 61  061 -0.5189073044  0.4450680683
#> 62  062  0.3744529500 -1.4569745257
#> 63  063 -0.3923133367  1.8857563126
#> 64  064 -0.5374822417 -0.9456266123
#> 65  065 -0.4123895122 -0.3262768273
#> 66  066  0.1905916121 -0.6597077729
#> 67  067  0.7631217926 -0.1566799579
#> 68  068 -0.1496676904 -0.5061193854
#> 69  069 -1.0884885816 -1.5529522231
#> 70  070 -1.1823723908  1.0657454024
#> 71  071  0.9383427374  0.6202569818
#> 72  072  1.7381321837  0.4611406675
#> 73  073  1.5165664257  0.9431975011
#> 74  074 -1.8151861844 -2.6800913824
#> 75  075  0.1118849893  0.6821351196
#> 76  076 -1.3704169028 -1.8324183886
#> 77  077 -0.3503548480 -1.0484274136
#> 78  078  1.0787708880 -0.0406503666
#> 79  079  0.8821538997  0.7843305975
#> 80  080 -0.0874785888  1.1980782746
#> 81  081 -0.6786554766  0.5090640685
#> 82  082  1.7194093005  0.3533651696
#> 83  083 -0.7616045815  0.2090725959
#> 84  084 -2.0499414728 -3.3104635195
#> 85  085 -1.2483684656  0.1372686352
#> 86  086  0.3311251560  0.6963014393
#> 87  087  1.8102883549  0.7130788773
#> 88  088 -0.3055005376 -0.4918491465
#> 89  089  1.5170564797 -0.0980266681
#> 90  090  1.0903077541  0.6569591631
#> 91  091  1.4715648742  0.8086604323
#> 92  092  0.2419140497 -0.4163308775
#> 93  093 -0.6616647159  1.3235476642
#> 94  094 -0.3304386672  0.3096565932
#> 95  095  1.3301557473  1.2119733581
#> 96  096  0.0005233303  0.1721178494
#> 97  097  0.8082899126  0.0399687915
#> 98  098 -0.0158517028 -0.0090387643
#> 99  099 -1.5148662949  0.0640985756
#> 100 100 -0.1224171117  0.4921421961