Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5064778
#> Y_2 0.5064778 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4532291 -0.5948141
#> 2 -1.4635784 -1.4683489
#> 3  1.2177526  0.3028820
#> 4 -1.1820939 -1.2131835
#> 5  0.8106237  0.5702941

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.02244668 -0.222821259
#> 2   002 -1.17342843 -0.925543241
#> 3   003 -0.29217596  0.089827292
#> 4   004  2.24892063  0.357444368
#> 5   005 -0.23190160 -0.078550336
#> 6   006  0.19420631  0.130556605
#> 7   007  0.02300409 -0.895885477
#> 8   008 -0.69132082 -1.813561147
#> 9   009 -0.38671022 -0.525299788
#> 10  010  0.40447209  0.710932742
#> 11  011 -1.94580443 -0.769700163
#> 12  012 -0.54720410 -1.277980667
#> 13  013  0.24993596  1.530918111
#> 14  014 -0.49919528 -0.520083824
#> 15  015 -2.55037552 -2.638990586
#> 16  016  0.10777330  1.625111771
#> 17  017  0.56579128 -0.580291701
#> 18  018  0.17694709 -0.040923231
#> 19  019 -0.89261731 -0.880755719
#> 20  020  0.60649331  0.369388532
#> 21  021 -0.09180981 -0.105213308
#> 22  022  0.67985989 -0.466698595
#> 23  023  0.49031045 -0.942562230
#> 24  024  1.15614275  0.430747376
#> 25  025  3.18415959  0.986056025
#> 26  026 -0.99730499  1.210658855
#> 27  027  3.17431298  2.810572566
#> 28  028  0.92612312 -0.027597280
#> 29  029  0.77435104  0.999844814
#> 30  030 -1.60348157 -0.547450375
#> 31  031  0.92756058  0.047527672
#> 32  032  0.62101466  0.039747194
#> 33  033  0.36905460  0.212210766
#> 34  034  0.12800270  0.290302462
#> 35  035  0.26469361  0.519730559
#> 36  036  0.21898517  0.974200006
#> 37  037  0.54031051 -0.042318874
#> 38  038 -0.44649003  0.319419270
#> 39  039  1.03292527 -0.488528577
#> 40  040  0.63165987 -0.446307060
#> 41  041  1.28829528  2.057345304
#> 42  042  1.21629017  0.352265423
#> 43  043  0.80781553 -0.579286440
#> 44  044  0.61713874  0.943895564
#> 45  045 -1.32566473 -0.742836358
#> 46  046 -0.38108040  2.285569957
#> 47  047  0.71090807  0.420390926
#> 48  048  0.65049324 -0.194772553
#> 49  049 -1.03598881 -0.001474774
#> 50  050  0.73754337 -1.072113309
#> 51  051 -1.43861122 -0.626600080
#> 52  052 -0.07277862  0.067099177
#> 53  053  1.34315700  2.078732863
#> 54  054 -0.62363018 -0.757000326
#> 55  055 -0.19167167  0.085344917
#> 56  056 -1.10578040  0.014228587
#> 57  057 -1.59157630 -3.073687001
#> 58  058 -0.53286323  1.222547429
#> 59  059  0.17449765 -0.563193284
#> 60  060  1.38173494 -0.289060691
#> 61  061 -0.11082609 -0.965249250
#> 62  062 -0.77211948  0.373595344
#> 63  063 -0.96679317 -0.870047226
#> 64  064 -1.43480125 -0.590919978
#> 65  065 -0.87440568 -0.569144280
#> 66  066 -0.59706363  0.670933689
#> 67  067  0.15829389  0.073106631
#> 68  068 -0.34999629  0.274825578
#> 69  069  0.50948191 -0.760966670
#> 70  070 -2.36399818 -2.414728816
#> 71  071  1.19293100 -0.722872360
#> 72  072  2.01345735 -0.852619830
#> 73  073 -1.15440964 -0.482058569
#> 74  074 -2.36940408 -1.417955164
#> 75  075 -1.44258381 -1.228133090
#> 76  076  1.28548865  0.787470347
#> 77  077 -0.31958296  0.074712157
#> 78  078  0.80052089  0.967904438
#> 79  079 -0.22475996 -0.167418828
#> 80  080 -0.72444695 -1.407345561
#> 81  081 -1.91224229 -1.242893196
#> 82  082  0.68414536  0.442446246
#> 83  083 -1.34981936 -0.784049726
#> 84  084  1.34080217 -0.396296887
#> 85  085  1.05536612  1.080879141
#> 86  086 -1.05703632  0.108306930
#> 87  087 -2.53031026 -0.957120639
#> 88  088 -0.19678391 -0.613996646
#> 89  089 -0.14004918  0.435128833
#> 90  090 -0.84422295 -1.588609669
#> 91  091 -1.68565059 -0.506580087
#> 92  092  0.34533230  0.371279090
#> 93  093 -2.39685301 -0.425034972
#> 94  094  0.77376821  0.040624905
#> 95  095  0.29351263  0.343818532
#> 96  096 -0.69315504 -1.128886404
#> 97  097 -0.10509858  0.976015409
#> 98  098  0.62344744  0.274712100
#> 99  099  0.89083993  1.215317887
#> 100 100  0.71583251  0.274280988

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.03255935  0.09877272
#> 2   002  1.33020903  0.34404757
#> 3   003  0.88047419  0.97431572
#> 4   004 -0.90873321 -0.06129678
#> 5   005  1.32355884  1.03394399
#> 6   006 -1.97090348 -2.27445321
#> 7   007  0.57046926 -0.04765198
#> 8   008 -0.41891693 -0.43407413
#> 9   009 -1.01836461 -1.79465454
#> 10  010 -0.06305145  0.46107881
#> 11  011  0.18628937 -0.18488174
#> 12  012 -0.47760535 -0.16487445
#> 13  013  0.99421865  1.27433983
#> 14  014 -1.21250974 -1.83216813
#> 15  015  0.51532650  1.40459116
#> 16  016  0.65838643  1.06561058
#> 17  017 -1.50414786 -0.33165436
#> 18  018 -0.25192048 -1.93938576
#> 19  019 -1.48421589 -0.99002595
#> 20  020  1.13110798  1.27716245
#> 21  021 -0.56785558  0.23158749
#> 22  022 -0.11416787 -0.52625952
#> 23  023  0.80246170 -0.05742406
#> 24  024 -1.58183903 -0.21760835
#> 25  025  1.49101221  0.74123574
#> 26  026 -0.12823840 -0.24267892
#> 27  027  1.57966723  1.95038778
#> 28  028 -0.57887976 -1.16878179
#> 29  029 -1.28913547 -0.31961933
#> 30  030  0.53936879  0.28621802
#> 31  031 -0.26252715  0.53180737
#> 32  032 -1.11221503 -0.55866638
#> 33  033 -1.68377204 -1.04760723
#> 34  034 -2.27593964 -0.97513901
#> 35  035 -0.21092099  1.48923581
#> 36  036 -0.60475265 -1.32020400
#> 37  037  0.79129607  0.83956034
#> 38  038 -1.59400698 -1.59897322
#> 39  039  0.52038258 -0.22166102
#> 40  040  1.59361512  1.06517427
#> 41  041  0.44768907 -0.24023024
#> 42  042  1.23090856  0.80316503
#> 43  043 -0.15359155 -1.98527375
#> 44  044 -2.41784865 -1.57171512
#> 45  045  0.51692468 -1.08578714
#> 46  046  1.01055497 -0.10435490
#> 47  047  1.21309767  1.22477251
#> 48  048 -0.19258010  0.02681984
#> 49  049  0.49034599  0.53899671
#> 50  050  0.83236748  1.50134997
#> 51  051 -0.05171936  1.40512830
#> 52  052  0.52181511  2.32262032
#> 53  053 -2.58205526 -1.98991583
#> 54  054  0.60027351 -1.01725748
#> 55  055 -0.07160297  0.77720384
#> 56  056 -0.07308149  1.95862617
#> 57  057  0.95514350  1.04470835
#> 58  058 -0.04136878  0.52385620
#> 59  059  0.16509360 -1.36523909
#> 60  060 -0.79739436 -0.14514130
#> 61  061  0.26460724 -0.33217700
#> 62  062  0.03553616  0.27210444
#> 63  063 -1.36818294 -0.09508299
#> 64  064  0.06782404  1.04876803
#> 65  065  0.91166324  1.06871582
#> 66  066  1.18643270  0.87738713
#> 67  067  0.41525151 -0.50557614
#> 68  068 -0.20842759 -0.20316949
#> 69  069  0.29600162  1.08091470
#> 70  070 -1.66822977 -1.00049193
#> 71  071 -0.74963554  0.26941011
#> 72  072  0.46186346 -0.22824480
#> 73  073 -1.18951832 -1.09175673
#> 74  074  1.91297357  0.73764423
#> 75  075 -1.41810299 -3.25393829
#> 76  076 -0.17262196 -0.26750459
#> 77  077  0.79207122 -0.61777678
#> 78  078 -0.74237521  0.95639753
#> 79  079  0.31096199  0.54233782
#> 80  080 -0.22468336  0.06799696
#> 81  081  0.13218251  0.81346172
#> 82  082 -0.06988263 -0.55457009
#> 83  083 -2.17719921 -1.08850342
#> 84  084 -0.67755057  0.04826643
#> 85  085  0.49411292  1.20859157
#> 86  086 -0.24040096 -0.73926614
#> 87  087 -2.04592508 -0.26868976
#> 88  088  0.22562101  0.54598442
#> 89  089 -0.22500329 -0.30707262
#> 90  090  0.04250823  0.56268697
#> 91  091 -0.80919812 -0.04181903
#> 92  092 -0.43313844  0.30869355
#> 93  093  0.71137632  1.23432413
#> 94  094 -1.23570762 -0.23156294
#> 95  095 -0.22454026 -0.57681573
#> 96  096  0.20734661  1.13137646
#> 97  097  0.67811424  0.75347484
#> 98  098  0.40156029 -0.64594196
#> 99  099  0.21392372  0.49950285
#> 100 100  0.27801932 -0.26509258