Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4701513
#> Y_2 0.4701513 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1961272  0.9017619
#> 2  0.6082665  0.4619942
#> 3 -1.5982672  0.1530656
#> 4 -1.1488774 -0.3337296
#> 5 -0.7093842 -1.3541297

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1           Y_2
#> 1   001 -0.74676893 -0.0586184623
#> 2   002  1.33520805  2.1265573845
#> 3   003  0.39147978  0.0407505950
#> 4   004 -0.95444153 -0.8531305275
#> 5   005  0.11031051  0.6068232664
#> 6   006  1.48969098 -0.8011152534
#> 7   007  1.25696120  0.8929005244
#> 8   008  0.78888798  0.5700891289
#> 9   009 -0.04197408  0.7868735931
#> 10  010 -0.03023970 -1.5579009339
#> 11  011 -0.52844613 -0.5372584532
#> 12  012 -0.63528503 -0.1217563957
#> 13  013  0.36033184 -1.6238407196
#> 14  014  2.05345524  2.3243557823
#> 15  015 -0.59540286  0.4188246839
#> 16  016 -0.94400188  0.1421906518
#> 17  017  0.14255332  1.1810045429
#> 18  018  0.04257766 -0.5160949057
#> 19  019  1.33452257  0.0003490908
#> 20  020 -0.54727970 -0.6063829300
#> 21  021 -1.77915096 -1.9687858262
#> 22  022  0.18127570 -0.8874339517
#> 23  023  0.26096315 -1.7383081786
#> 24  024  0.72307219 -0.0891135051
#> 25  025  1.67535861  1.0464350028
#> 26  026  0.03749700  1.4526081321
#> 27  027  0.04922290 -0.8388553125
#> 28  028  1.45482869  0.3255028621
#> 29  029  1.51272771  1.0662132309
#> 30  030  0.81236179 -0.7410517362
#> 31  031  1.93372364  1.9574595001
#> 32  032  1.72197997  0.9608859665
#> 33  033 -0.08715692 -0.5129684213
#> 34  034 -1.67142543 -0.9053381549
#> 35  035 -0.86497276 -0.3401455379
#> 36  036  1.24124657  2.0899950639
#> 37  037 -0.67141419 -1.0665987341
#> 38  038 -1.40433914 -2.5035696517
#> 39  039  1.78895124 -0.2923967723
#> 40  040 -1.06692524 -1.5819924282
#> 41  041 -1.88096029 -1.0224671934
#> 42  042  0.51803900  0.2182202472
#> 43  043  1.33120487  1.0876464766
#> 44  044  0.69635215  0.0699778485
#> 45  045  0.88624660 -0.1825188873
#> 46  046  0.81869747  1.3228135977
#> 47  047  0.84118237 -0.2520591146
#> 48  048 -0.36117104  0.2124189733
#> 49  049 -0.25390939 -0.5249021029
#> 50  050  1.30317610  0.3267511110
#> 51  051  2.15594135  0.3842577129
#> 52  052 -0.02141469 -1.7418642156
#> 53  053  0.09189012 -1.3180232698
#> 54  054  0.66037167  0.2154825357
#> 55  055 -0.19580484 -0.8400315988
#> 56  056 -0.70343316 -0.8807809695
#> 57  057  1.28239563  0.1703127468
#> 58  058 -0.09895352  0.6626756225
#> 59  059 -0.01164194 -0.2250977964
#> 60  060  0.18752370 -0.1376319784
#> 61  061  0.18863628 -1.1146829473
#> 62  062 -0.61413588 -0.5528071184
#> 63  063  0.68211481 -0.2587032204
#> 64  064 -0.45278297 -0.5585158528
#> 65  065  1.31536546  1.5368568357
#> 66  066  0.25127422  1.1330071874
#> 67  067 -0.12770952  0.5717860653
#> 68  068  0.60821541  0.2683811288
#> 69  069  0.83553930 -0.0535700176
#> 70  070 -0.05224119 -0.8954362225
#> 71  071  1.25073350 -0.4145007605
#> 72  072 -1.25550263 -1.9015167487
#> 73  073 -0.40038871  1.1368977495
#> 74  074  1.07790946  0.9169977780
#> 75  075 -0.32488651 -0.7283795818
#> 76  076 -0.44839611  0.2635431460
#> 77  077  1.08108221  0.4358315712
#> 78  078 -0.23504552 -0.0907752714
#> 79  079  0.39545606  1.2239797484
#> 80  080 -2.89514752 -1.8226982523
#> 81  081  0.11435387  0.0584351458
#> 82  082  0.21507517  0.9317756212
#> 83  083 -1.78805765 -1.0002565808
#> 84  084  0.70452844 -0.8114420954
#> 85  085 -0.19021172  1.3112082938
#> 86  086 -1.23869343  0.1349768281
#> 87  087 -0.70005371 -1.2716302979
#> 88  088 -0.41688210 -0.3729438908
#> 89  089 -1.57779474 -1.8329723080
#> 90  090 -0.18447238 -0.5017756028
#> 91  091  1.28412286  0.0739551772
#> 92  092  1.34074996  1.5764993313
#> 93  093 -0.80361626 -0.5493143021
#> 94  094  0.20919388 -0.2468038103
#> 95  095 -0.09058890  1.5567491773
#> 96  096  1.34596481  0.8571710638
#> 97  097  1.75086861  1.0632100989
#> 98  098  0.49616802  0.0665260568
#> 99  099  1.38667037  3.4466652733
#> 100 100  0.22553411  0.5692551955

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  1.036131751  0.327124599
#> 2   002 -1.482927715  0.008167065
#> 3   003 -0.635997608 -0.692304073
#> 4   004  1.150199463 -0.244226553
#> 5   005  1.785526014  1.857571161
#> 6   006 -1.055346416  0.614006458
#> 7   007  0.920606015  1.398225579
#> 8   008  0.307353480  0.292576373
#> 9   009 -1.350779543 -2.025435542
#> 10  010  1.954549575  1.230786744
#> 11  011 -1.016384110 -2.231527234
#> 12  012 -2.100353327 -1.758569771
#> 13  013  0.918459032 -1.657289445
#> 14  014 -1.194577490 -0.611231526
#> 15  015 -2.003976577  0.122822203
#> 16  016  1.470309956  0.732349979
#> 17  017  0.645589238 -0.639222928
#> 18  018  0.089482588  0.350614845
#> 19  019 -0.003786432 -0.477253501
#> 20  020  0.965520251  0.190651908
#> 21  021 -2.528881215 -1.796714344
#> 22  022  0.561692737  2.328017560
#> 23  023 -0.778108605 -1.626336015
#> 24  024 -0.788418794 -0.449409533
#> 25  025 -0.326737742  0.301751123
#> 26  026 -1.064370306 -0.869599123
#> 27  027 -0.079912623  0.807370008
#> 28  028  0.229624331  1.237353348
#> 29  029 -2.136081676  0.135185921
#> 30  030 -0.866263696  0.121361392
#> 31  031 -2.721982883 -1.232208888
#> 32  032  0.761266739 -0.182587706
#> 33  033 -1.494675941 -0.113028243
#> 34  034 -3.071483448 -2.757734456
#> 35  035 -0.763129915 -1.076698720
#> 36  036  0.086682708 -0.302058494
#> 37  037  0.088395929 -0.726786939
#> 38  038 -0.455849488  1.107946046
#> 39  039  0.870510927  1.793933229
#> 40  040  0.509503388  0.829242164
#> 41  041 -1.921769040 -0.818492850
#> 42  042 -1.415504555  0.580797797
#> 43  043  0.438187931  0.471680148
#> 44  044  0.162692791 -0.049985019
#> 45  045 -0.462047349  0.927576271
#> 46  046  1.932291421 -0.340304519
#> 47  047 -0.237152315  0.291065703
#> 48  048 -0.434027257 -0.483964971
#> 49  049 -0.598684657 -0.163495416
#> 50  050 -1.529630899 -1.060697213
#> 51  051  0.153653756  1.154032610
#> 52  052  0.484561833 -0.086495853
#> 53  053 -1.439000987 -0.905020301
#> 54  054 -1.241281854  0.786920891
#> 55  055 -0.473458213 -1.510776233
#> 56  056 -0.719652920 -0.847788947
#> 57  057 -0.778029919 -1.566375064
#> 58  058 -0.487838111  0.144999198
#> 59  059  0.339292670  0.501488269
#> 60  060  0.668763762  0.542377299
#> 61  061 -1.285279530 -2.019692294
#> 62  062  0.801017502  0.844702451
#> 63  063 -0.140795023  0.183420852
#> 64  064 -0.826696762 -1.744206369
#> 65  065 -0.839420191 -1.024186849
#> 66  066  0.958315321 -0.864180797
#> 67  067  0.397081791  1.518721824
#> 68  068  0.336496344 -0.073877383
#> 69  069 -1.067313127  0.155094077
#> 70  070  0.487128700 -0.241101764
#> 71  071  0.519681090  1.180465863
#> 72  072  0.204672968 -0.823429140
#> 73  073  0.881315387  0.943878288
#> 74  074  0.797900117 -0.072709254
#> 75  075 -0.763271389 -1.613600645
#> 76  076  0.636235380  0.072345182
#> 77  077  0.671243037  0.190104909
#> 78  078  0.642941721 -0.660471900
#> 79  079 -0.841553474 -0.551426337
#> 80  080  0.943235865 -0.875176065
#> 81  081 -1.177163649 -0.904796596
#> 82  082 -1.236062643  0.495827896
#> 83  083  0.916247100  0.089483434
#> 84  084 -2.767095316 -0.846283904
#> 85  085  0.073513729  0.050039372
#> 86  086 -1.341529213 -2.451118915
#> 87  087 -0.572048402  0.657405550
#> 88  088  0.394361506  1.081504859
#> 89  089 -0.533983526 -1.021727905
#> 90  090  2.654233940  1.372592974
#> 91  091  0.371154224  0.962422617
#> 92  092  0.602978107  0.900665961
#> 93  093  0.166791065  1.159834002
#> 94  094 -0.440135939  1.050128519
#> 95  095  0.629298578  1.392072647
#> 96  096  0.886108581  2.104550167
#> 97  097 -0.414956236  0.042476008
#> 98  098  2.701906075  1.388328002
#> 99  099 -2.757222657 -1.223247884
#> 100 100  0.422105422 -0.409570737