Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5446491
#> Y_2 0.5446491 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4844384  0.4065709
#> 2  0.3230008 -1.0654424
#> 3 -1.6295140 -1.4937146
#> 4  0.8901652 -0.3696356
#> 5  0.2813043  1.1167436

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.966694763  0.321592072
#> 2   002  0.560202357 -0.216766525
#> 3   003  0.511666106  0.905981925
#> 4   004 -1.129703688 -1.302135545
#> 5   005 -1.125963879  0.352200802
#> 6   006  2.311098281  1.760319257
#> 7   007  1.289577970  0.749817773
#> 8   008 -0.228123320  1.070936256
#> 9   009 -1.245149786 -0.053826607
#> 10  010  1.760151111  0.899040862
#> 11  011  0.513104885  0.142311416
#> 12  012  0.039637958  0.294422452
#> 13  013  1.029649652  1.039242048
#> 14  014 -0.327394901 -0.268189223
#> 15  015  0.477238119 -0.300216362
#> 16  016  0.417880525  0.522888203
#> 17  017 -0.889108587 -2.044210933
#> 18  018  1.340801373 -0.268827166
#> 19  019  1.613183195  1.204167291
#> 20  020 -0.358704315 -1.777753475
#> 21  021  0.320870591 -0.767775579
#> 22  022 -1.389555899 -1.658856147
#> 23  023  0.506518502  0.167521564
#> 24  024 -0.323820408 -0.702228761
#> 25  025 -0.316678425  0.326589508
#> 26  026  0.221990299  1.209755295
#> 27  027 -0.232535641 -1.228529117
#> 28  028  1.995642466  0.326980495
#> 29  029 -0.888716002  0.176294233
#> 30  030  0.886646137  1.027438536
#> 31  031  0.567989241 -0.062247548
#> 32  032  0.022892264  0.965420007
#> 33  033  0.559903353  0.341320076
#> 34  034  1.584481697  0.599497681
#> 35  035  1.284447712 -0.174235127
#> 36  036  2.301486368 -0.044432863
#> 37  037 -1.450283321 -1.471131923
#> 38  038  0.520700288  1.270198437
#> 39  039  0.759431739  0.990814068
#> 40  040  0.497129967 -0.171740674
#> 41  041  0.935252748  0.557054328
#> 42  042  1.782045918 -0.064611355
#> 43  043  0.562617492 -0.250771386
#> 44  044  0.996538391 -0.157076631
#> 45  045 -0.450623847 -0.933530389
#> 46  046 -0.009941707 -1.310513987
#> 47  047 -1.940965168 -1.283821313
#> 48  048 -0.713057274 -0.894138337
#> 49  049  0.502250047 -0.442844465
#> 50  050 -0.321438773  1.811131173
#> 51  051  0.248878057  0.144226400
#> 52  052  0.673111661  0.696623065
#> 53  053 -1.896251559  1.171466950
#> 54  054 -1.504404394  0.772278241
#> 55  055 -1.941454262  1.003828980
#> 56  056 -0.836769119 -1.302991687
#> 57  057 -1.538438844 -0.670026176
#> 58  058 -0.778063344 -1.453640746
#> 59  059 -1.167746399  0.320935911
#> 60  060 -1.211111901 -0.282182613
#> 61  061  0.171573968  0.197488281
#> 62  062 -0.206380684 -1.210147578
#> 63  063 -0.882338530 -1.287442311
#> 64  064 -0.500952315  0.396814190
#> 65  065  0.115293160 -0.227708702
#> 66  066  1.204561551  0.587815412
#> 67  067 -0.636537744 -1.055238541
#> 68  068 -0.084859764 -0.335960645
#> 69  069 -2.239052596 -1.659097743
#> 70  070  1.204666502  1.762058555
#> 71  071  0.108821421 -0.001379322
#> 72  072 -1.600345814 -1.037258493
#> 73  073  0.618470198 -0.485176508
#> 74  074  0.345079064  0.022814810
#> 75  075  0.250895250  0.929209408
#> 76  076  0.494546341  0.662963224
#> 77  077  0.108231913  0.266003607
#> 78  078 -1.311269394 -0.670397182
#> 79  079 -1.424835560 -0.486372549
#> 80  080  1.793925928  2.540475179
#> 81  081 -1.996380003 -0.879374320
#> 82  082 -0.256610379  1.280838291
#> 83  083 -0.045494431 -0.084545017
#> 84  084 -0.270436778 -1.895251725
#> 85  085  0.039203181  0.026577238
#> 86  086  1.018606675  0.060654553
#> 87  087 -0.187074713  1.692765301
#> 88  088 -0.049146721  1.119182675
#> 89  089  1.931330659  1.731584497
#> 90  090  2.104564656  1.245356853
#> 91  091  0.379405216 -1.528808798
#> 92  092 -0.514906311 -2.187497423
#> 93  093  0.660988503  0.109881227
#> 94  094 -0.606541761 -1.043888556
#> 95  095 -0.562239010 -0.860111570
#> 96  096  2.201830821  1.037950920
#> 97  097 -0.234952499  0.240469427
#> 98  098  0.450051918  0.408574479
#> 99  099 -1.287829843  0.980234844
#> 100 100 -1.042474228 -1.582841594

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1           Y.2
#> 1   001 -0.123485587  0.7116243775
#> 2   002  0.006104058  1.1291536646
#> 3   003  1.300966651 -0.7112089208
#> 4   004  1.319432180 -0.5512975252
#> 5   005 -0.937631485 -0.7739752986
#> 6   006 -0.494865767  0.3109812436
#> 7   007  1.090013866  1.1947026887
#> 8   008  1.324070025  1.4165397321
#> 9   009  0.746648679  0.3076414502
#> 10  010 -0.260170575 -0.4205404473
#> 11  011  0.217222940 -0.7394751358
#> 12  012 -0.460428167  0.0816362990
#> 13  013  1.363634899 -0.4857985184
#> 14  014  0.881443120  2.0279188563
#> 15  015 -0.250280639  1.1895947179
#> 16  016 -0.559722544 -0.1594548076
#> 17  017  0.320226206 -1.3360321967
#> 18  018 -1.766609060 -0.5638114712
#> 19  019  0.543117850 -0.5452856213
#> 20  020  0.007859133  0.9708049260
#> 21  021 -0.539449895 -1.1580375050
#> 22  022  0.201376238  0.5027894362
#> 23  023  0.349967973 -0.3368849336
#> 24  024 -0.584077676 -0.1824227804
#> 25  025 -0.132109714 -1.3176354198
#> 26  026 -0.326450821  0.4167800577
#> 27  027  0.569077878  0.5929953543
#> 28  028  0.378106490  0.3493733756
#> 29  029 -2.569462698 -2.3405173787
#> 30  030 -1.608946261 -1.9227928720
#> 31  031 -0.666601815 -0.8786606127
#> 32  032  0.179147033  0.6273393403
#> 33  033 -0.317534918 -1.2757727998
#> 34  034  0.313025842  1.3035721998
#> 35  035 -0.345291328 -0.2349044451
#> 36  036  0.390833084  0.1673919911
#> 37  037  0.904000845  0.0623820918
#> 38  038  0.334974789  0.5629391484
#> 39  039 -0.429906859 -1.0655265367
#> 40  040 -1.064042022  0.3183608013
#> 41  041 -0.991279249 -0.6922946446
#> 42  042  1.727934169 -0.9003897969
#> 43  043  0.112808023  0.1275970958
#> 44  044 -0.753916540  0.3176422700
#> 45  045  0.652590815  1.6104055117
#> 46  046 -1.136865140 -0.3883906592
#> 47  047  0.310997161  1.2267574481
#> 48  048  0.417712743  0.3727756248
#> 49  049  0.575544608  0.8752988594
#> 50  050 -0.408102393  1.1870032465
#> 51  051 -0.178155576 -3.1385801454
#> 52  052 -1.018866916  0.4238921924
#> 53  053 -0.083407234 -0.9876278099
#> 54  054  0.895634245  1.1330380882
#> 55  055 -2.278542021 -2.5162677008
#> 56  056 -0.018940474  0.7206268362
#> 57  057 -0.732868220 -0.0004754209
#> 58  058  2.044739184  0.1052461433
#> 59  059  0.478798438  0.4575161023
#> 60  060 -0.028253650  0.3108870831
#> 61  061 -0.972662244 -0.7431904346
#> 62  062 -0.741726718 -0.8873854229
#> 63  063 -0.751297969 -1.0811052802
#> 64  064 -0.790859235  0.6789243537
#> 65  065 -0.027278462 -0.5745972026
#> 66  066  0.990149227  2.5647940482
#> 67  067  1.497848175  0.5169972247
#> 68  068  0.350821760  0.0482138976
#> 69  069  0.355065820 -0.1066054352
#> 70  070  0.330024197  0.4751687796
#> 71  071 -0.160457946  1.2440964792
#> 72  072 -0.057937887 -1.7524706816
#> 73  073 -0.014551573  0.8080780785
#> 74  074  0.764432385 -0.0570705063
#> 75  075  2.018939323  1.9342325187
#> 76  076  0.305067280  0.3138211658
#> 77  077 -0.756702397 -1.9468513912
#> 78  078  0.023553811  1.5485269858
#> 79  079  0.453890449  0.1841937926
#> 80  080  1.479870749  0.6133510650
#> 81  081 -0.665371048  0.3528617143
#> 82  082  0.733664994 -0.5025635357
#> 83  083 -0.763698757  1.4003781322
#> 84  084  1.367429495  0.3183318632
#> 85  085 -1.622698581 -0.6123124881
#> 86  086  0.227395022 -0.9809068001
#> 87  087 -0.442468877 -1.3827947281
#> 88  088  0.760921315 -1.0020061929
#> 89  089 -0.642947461 -0.5997293884
#> 90  090 -1.063626937 -0.0897457404
#> 91  091  1.676915849 -0.5136957436
#> 92  092 -0.583552542 -0.9027914825
#> 93  093 -0.959265372 -0.0906653498
#> 94  094  1.680776137  2.0166729235
#> 95  095 -1.044369498  0.0814708614
#> 96  096 -1.769619016 -1.7520730230
#> 97  097  0.041194488 -0.6912976707
#> 98  098 -1.562006070 -0.8940843541
#> 99  099 -0.174731084 -0.4173358323
#> 100 100  1.110065720  1.4990750716