Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5116769
#> Y_2 0.5116769 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.1760452 -1.25568557
#> 2  0.3508308 -0.60022885
#> 3 -0.4894695 -0.78796002
#> 4 -1.1661118  0.01517764
#> 5 -1.0229721 -1.21375657

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1         Y_2
#> 1   001 -1.0371990028 -1.16747393
#> 2   002 -0.1983717261  0.69198627
#> 3   003  0.0861520434  0.32565070
#> 4   004  1.3435501973  1.54461747
#> 5   005 -0.1756708411  0.12716443
#> 6   006  0.4850655955  1.32586546
#> 7   007 -1.2497880624 -0.27275028
#> 8   008 -1.0657801357  0.05119155
#> 9   009  1.0249090906  0.77707266
#> 10  010 -0.6473435564 -0.89870169
#> 11  011 -0.2498766280  0.18951235
#> 12  012 -0.3560776356 -1.63424128
#> 13  013  0.0879387800  0.11330518
#> 14  014 -2.7896684144 -1.69061920
#> 15  015  0.2572560105  1.03819063
#> 16  016 -1.0718747115 -1.88059730
#> 17  017  2.3350068866  0.85825142
#> 18  018  1.4263716905  1.65708051
#> 19  019  0.9273073420  0.21065380
#> 20  020  1.6816825624 -0.76577003
#> 21  021  1.0570819831  1.53655682
#> 22  022  1.0858259366  0.86774351
#> 23  023 -1.4538322707 -0.68901889
#> 24  024 -1.4784902350 -0.98531741
#> 25  025  0.8024055755  1.28114733
#> 26  026  0.7355417574 -0.56594858
#> 27  027 -0.8023239757  0.78346320
#> 28  028 -0.4261368760  0.69330354
#> 29  029 -0.3974081991 -0.45215060
#> 30  030 -0.6981544588 -0.90163095
#> 31  031 -1.2965714058 -0.87298584
#> 32  032 -0.1062313433  1.31001353
#> 33  033 -2.0136413373 -1.76047287
#> 34  034  0.1964107875  1.17285682
#> 35  035 -0.4691406744  0.55208561
#> 36  036  0.5290333003 -0.91462415
#> 37  037 -0.1160794278  0.46261155
#> 38  038  0.0954712550  0.74287249
#> 39  039 -2.8192182254 -3.63567490
#> 40  040  0.2422167830 -0.50739408
#> 41  041  0.5953863730  0.55985861
#> 42  042  1.3417241449  0.52491819
#> 43  043 -0.1088587322  0.67670244
#> 44  044  0.5683054371  1.62281591
#> 45  045  0.1820582513  0.05188170
#> 46  046 -0.4485229579 -1.56506862
#> 47  047  0.1172190117 -0.74908882
#> 48  048 -2.0659942958 -1.55281395
#> 49  049  0.6386975872  1.66466053
#> 50  050 -0.6011595062 -0.12878702
#> 51  051 -1.5041163627 -0.09197817
#> 52  052  1.1146187727  0.58160533
#> 53  053  0.6708532722  1.08924752
#> 54  054  0.4931734387 -0.98966130
#> 55  055 -0.0799916833 -0.48590995
#> 56  056 -0.1124225785  0.10302029
#> 57  057 -0.6889941542 -1.31401633
#> 58  058 -0.9487262763  0.21737327
#> 59  059 -0.0981578563 -0.11784212
#> 60  060 -0.4948896382 -0.41323018
#> 61  061 -0.2796630846  0.15753154
#> 62  062 -2.0399959851 -0.46125161
#> 63  063  0.3663204557 -0.30047728
#> 64  064  0.6521315915  0.36709868
#> 65  065  0.2224414049  0.77141604
#> 66  066 -0.2513130933 -0.34729039
#> 67  067 -1.1009216679 -0.19231985
#> 68  068  0.4454809740 -0.68523350
#> 69  069  0.4091315376 -0.59490504
#> 70  070  0.5016106318  0.50581513
#> 71  071 -0.9927684973 -1.53466239
#> 72  072  2.1345770632  0.60505080
#> 73  073  0.3995926420 -0.06988458
#> 74  074 -0.0009093057  0.13945860
#> 75  075  0.3995400895 -0.29748106
#> 76  076  0.6742755073  0.48590227
#> 77  077 -1.7477071475  0.03792132
#> 78  078  0.9616142354  0.62262708
#> 79  079 -0.8335591745  0.51151183
#> 80  080 -0.3282991505 -0.16988007
#> 81  081 -0.5706134618 -1.01120987
#> 82  082  0.8130150799  0.21546940
#> 83  083  0.2091488402 -0.10630036
#> 84  084  0.4198482199 -0.94337844
#> 85  085  0.4062264732  0.19703786
#> 86  086 -0.0818427012  0.42200912
#> 87  087 -0.1045447933 -0.71592692
#> 88  088 -0.9638330264  1.10551719
#> 89  089 -0.2879660607  0.03184028
#> 90  090 -0.0479469098  0.65878270
#> 91  091 -0.7560864365 -0.97562451
#> 92  092  1.4186286317  0.61942163
#> 93  093 -1.8359973311 -1.35802553
#> 94  094 -1.4750447377 -0.10511770
#> 95  095  1.8804962381 -1.05208886
#> 96  096 -2.0066269589 -2.16123694
#> 97  097  0.1965827725 -0.21348548
#> 98  098  0.8769336200  0.47844479
#> 99  099 -1.4005178105 -1.30506755
#> 100 100  1.0050498797 -0.96370120

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.026482298 -0.421095422
#> 2   002  0.510355455  0.500551891
#> 3   003 -1.663317304 -1.099732537
#> 4   004  0.171829118 -0.059019530
#> 5   005 -1.055502717 -1.130532637
#> 6   006 -0.051041624  0.310689187
#> 7   007  0.728288175  0.829427709
#> 8   008  1.754191752  1.971964781
#> 9   009  0.234336896  0.257526746
#> 10  010 -1.191711469 -0.152020569
#> 11  011  1.473379840  1.294794086
#> 12  012  2.174606660  1.094077093
#> 13  013 -0.857860337  1.226705295
#> 14  014  0.007545099  0.327012636
#> 15  015 -0.952210930 -0.632491204
#> 16  016  1.693678246  1.914260486
#> 17  017  1.124407777  1.148462089
#> 18  018  0.778229783  1.101400376
#> 19  019  0.230808578 -1.506358722
#> 20  020  0.796144021 -0.226052206
#> 21  021 -0.461335784 -2.190317335
#> 22  022 -2.455079699  0.918539362
#> 23  023 -0.741540741 -1.063401935
#> 24  024  0.759273055  1.878223011
#> 25  025 -1.286496307  0.214414320
#> 26  026  0.419859921  0.195683024
#> 27  027  0.713071653  3.129661814
#> 28  028  1.280483372  0.548544037
#> 29  029 -1.768806806 -0.880604760
#> 30  030  1.067660406 -0.538089882
#> 31  031  0.564680906  0.433700733
#> 32  032  0.225133906  0.003476889
#> 33  033  0.609570417 -1.231721342
#> 34  034  0.940377547 -1.224265042
#> 35  035  0.123783049  0.620159419
#> 36  036 -0.940633166 -0.137669431
#> 37  037 -0.443592482 -1.336658777
#> 38  038 -0.132922096 -0.961838058
#> 39  039 -0.129425126 -0.932690046
#> 40  040  0.613073063 -0.606574149
#> 41  041 -0.535988233 -0.153471696
#> 42  042  0.119222451 -1.317990515
#> 43  043 -1.778953619  0.490501148
#> 44  044  0.606862145  0.795975191
#> 45  045 -1.733435231 -0.698262689
#> 46  046  0.139979843 -0.210071366
#> 47  047  0.887909353  0.799071075
#> 48  048  0.247912183  0.479308582
#> 49  049 -0.087500308  0.331830540
#> 50  050  1.694129586 -1.293611897
#> 51  051 -1.129866701  0.118296295
#> 52  052 -0.169077153 -0.273486436
#> 53  053 -0.848466849 -0.496913708
#> 54  054  0.130971010  0.176985089
#> 55  055 -0.527427299 -1.850874952
#> 56  056 -0.185668899  0.236486415
#> 57  057 -2.007016325 -2.848612386
#> 58  058  0.770437624 -0.092821820
#> 59  059 -0.200671752 -0.607895158
#> 60  060  1.847830372  0.735058633
#> 61  061  1.377042433 -0.262659794
#> 62  062  0.681310149 -0.929295562
#> 63  063 -1.253949697 -0.509906507
#> 64  064  0.041008517  0.567735773
#> 65  065 -0.472913018 -0.343279585
#> 66  066  0.674169653  0.244410175
#> 67  067  2.542706799  3.270755493
#> 68  068 -0.766358478  1.419566533
#> 69  069  0.577309758 -0.324309641
#> 70  070  0.019446002 -0.632060243
#> 71  071  1.218236059  0.748420142
#> 72  072  1.408639688 -0.210093013
#> 73  073 -0.380622414 -0.049546393
#> 74  074 -0.358232983 -1.031128983
#> 75  075  0.939603203  1.182906694
#> 76  076  0.098132975 -0.109001353
#> 77  077  1.455757221 -0.553535212
#> 78  078  0.841284541  0.868926393
#> 79  079  1.650694099  0.856087478
#> 80  080 -0.375981970  0.617907327
#> 81  081  1.634191148 -0.199667871
#> 82  082  1.214301317  0.889392812
#> 83  083  0.446299764 -0.251918716
#> 84  084  0.850767141 -0.617840244
#> 85  085 -0.602227891  0.326039055
#> 86  086 -0.178759624  1.977079530
#> 87  087  1.415826858  0.128989195
#> 88  088 -1.268522548  0.804010742
#> 89  089 -0.455636376  1.619536056
#> 90  090  0.507320485  0.088102708
#> 91  091  1.237560640  0.900708986
#> 92  092  0.111853299  0.875660161
#> 93  093  0.166717501  0.853322845
#> 94  094 -0.772707830  1.229019339
#> 95  095 -0.236649455 -1.011383161
#> 96  096 -0.084881153  1.373933511
#> 97  097  0.351749134 -0.185713137
#> 98  098 -0.238424349 -1.695179406
#> 99  099 -0.054635462 -0.857221553
#> 100 100  0.057903414  0.577372118