Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5344037
#> Y_2 0.5344037 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1       Y_2
#> 1  1.01749029 2.4521770
#> 2 -0.47563438 0.8678445
#> 3  0.93357944 1.4250843
#> 4 -0.05192786 0.5516120
#> 5  0.88554232 0.9109807

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.87904479 -1.800093475
#> 2   002 -0.68270497 -1.044746089
#> 3   003 -0.01065854 -1.027996661
#> 4   004  0.08137209 -0.177555774
#> 5   005  1.10848694  1.270925223
#> 6   006 -0.59248971 -0.687095469
#> 7   007  1.29987681  0.522767607
#> 8   008  1.18726191 -0.709170526
#> 9   009  0.11344596  0.348454596
#> 10  010 -0.26015207 -0.252628313
#> 11  011  2.68532265  1.267338773
#> 12  012 -0.05142191  1.177399756
#> 13  013 -0.46596630  0.325280397
#> 14  014 -0.14381507  0.579059576
#> 15  015 -1.62233880  0.574671568
#> 16  016  0.33288962 -0.585375325
#> 17  017 -0.76765639 -0.633283514
#> 18  018 -0.07240776  1.574196972
#> 19  019  0.64591411  2.663473673
#> 20  020  0.12089362 -0.317796422
#> 21  021  0.32963451  0.405122982
#> 22  022  1.07001590  0.409414886
#> 23  023  0.58833994  0.215738844
#> 24  024  1.14092010  0.215117917
#> 25  025 -0.07129538 -0.246958604
#> 26  026 -0.16339394 -0.375183477
#> 27  027  0.22207610  0.360638795
#> 28  028  0.54244906 -0.984552245
#> 29  029  0.23737566 -0.062293839
#> 30  030  0.09472311 -0.796096859
#> 31  031 -0.18548692  0.177512671
#> 32  032  1.21297481  0.250534470
#> 33  033 -0.12871254  1.011995461
#> 34  034 -1.04992939 -0.780024242
#> 35  035  1.12672390 -0.004525884
#> 36  036  0.08249769  0.292917162
#> 37  037 -0.43562964  0.460158532
#> 38  038 -0.89305050 -0.165537227
#> 39  039 -0.15591825  1.056519342
#> 40  040 -1.06477029  0.063896309
#> 41  041  0.04782732  0.616292053
#> 42  042 -0.01830278  0.824543132
#> 43  043  0.84895920  0.014197263
#> 44  044 -0.74371388  0.469552916
#> 45  045  0.31584997  0.920213509
#> 46  046 -1.07113889 -1.871588951
#> 47  047 -1.38018554 -2.635561716
#> 48  048 -0.59181250  0.001934910
#> 49  049  1.00662604  0.325028867
#> 50  050 -0.28123802  0.974357956
#> 51  051  0.72442993 -0.341264028
#> 52  052  0.98908105  0.994889964
#> 53  053  2.08456504  0.595464132
#> 54  054  0.87448643  0.216314339
#> 55  055  0.41479524  1.057695851
#> 56  056  1.02908078  0.505117460
#> 57  057 -1.22660672 -1.058299358
#> 58  058 -0.04216461  0.899754108
#> 59  059  0.67712847 -0.037664755
#> 60  060 -0.10374810 -0.142610981
#> 61  061 -0.45142465  1.340583578
#> 62  062  1.77895416  0.828509695
#> 63  063  0.86702898  0.394567845
#> 64  064  0.82147915 -2.159246562
#> 65  065 -1.44934392  0.149804687
#> 66  066  1.03903778  1.188593592
#> 67  067 -0.45040353 -0.298818084
#> 68  068 -1.51834595  0.191337474
#> 69  069 -0.27859359 -0.316106029
#> 70  070  0.97800042 -0.255091244
#> 71  071  0.15109651  0.853330356
#> 72  072  0.33645408 -0.016861408
#> 73  073 -3.05508331 -1.537761827
#> 74  074 -2.05272856 -1.532149792
#> 75  075  0.33313356 -1.660117301
#> 76  076  0.71064809  0.557090412
#> 77  077  1.13529249  1.085298041
#> 78  078 -0.93319500 -0.089921044
#> 79  079 -0.05796824  0.708051620
#> 80  080  0.28309059 -1.050782154
#> 81  081 -0.07526918 -2.031793060
#> 82  082  0.26228827  1.182627587
#> 83  083  1.14973020 -1.063239897
#> 84  084  0.41139356  0.113180935
#> 85  085  0.89018308 -1.227962583
#> 86  086  2.52013086  2.356908773
#> 87  087 -0.86031279 -0.931843055
#> 88  088 -2.50907067 -0.583232254
#> 89  089 -1.62395362 -0.073141446
#> 90  090 -1.72949172  0.109901456
#> 91  091  0.49751238 -1.161008832
#> 92  092 -1.30764918  0.084092327
#> 93  093 -0.68590130 -1.219182507
#> 94  094 -0.60014690 -0.196903693
#> 95  095 -0.80813112 -2.116630570
#> 96  096  0.69020739  1.537708510
#> 97  097 -0.89500945  0.444165574
#> 98  098 -0.60224920  0.918441987
#> 99  099  1.57576566  0.522858324
#> 100 100 -0.43451959 -0.009519279

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.837501260 -0.68585713
#> 2   002  1.060384423  0.79449712
#> 3   003  1.011192615  0.02193318
#> 4   004  0.493289782 -1.15429594
#> 5   005  0.079303743  0.62696656
#> 6   006 -0.192862272 -0.14620877
#> 7   007  1.482867185  2.02454697
#> 8   008  0.578522531 -0.77271356
#> 9   009  1.939980799  2.74964411
#> 10  010 -0.136459537  0.04349957
#> 11  011  2.058417006  2.01290458
#> 12  012 -0.447483533 -0.19789974
#> 13  013 -0.167818032  0.72555210
#> 14  014 -0.115583405  1.40442453
#> 15  015  1.273771928  1.07536730
#> 16  016  0.341696722  0.88536098
#> 17  017 -1.714210984 -1.28143821
#> 18  018  1.228901997  1.03561508
#> 19  019 -0.007965999 -1.37347061
#> 20  020 -3.003607504 -2.64313198
#> 21  021 -0.739376938 -0.66779335
#> 22  022 -0.817287889 -0.63156571
#> 23  023 -1.711041999 -1.37137489
#> 24  024  1.792240425  0.37690155
#> 25  025 -0.157634892 -1.65588331
#> 26  026 -0.570619825 -0.69526837
#> 27  027 -0.680523945 -0.15718208
#> 28  028 -0.360699964 -0.67975494
#> 29  029 -0.319223166 -1.37686017
#> 30  030  0.012746917 -0.92325457
#> 31  031  0.243532639  1.16846837
#> 32  032  0.891652629  1.77641608
#> 33  033  1.432375223  1.09529585
#> 34  034 -0.121116144 -0.58151268
#> 35  035  0.772478749  0.35117425
#> 36  036 -1.061268873 -0.18864232
#> 37  037  0.384289208  1.12683267
#> 38  038 -0.393031823 -0.10926803
#> 39  039 -1.274682484  0.78720464
#> 40  040 -0.058774150  0.58030256
#> 41  041 -1.379665330 -0.38083049
#> 42  042 -0.235725124 -0.18258517
#> 43  043  0.583835164  1.11006500
#> 44  044  1.786896029  0.63541491
#> 45  045  0.843678688  1.14703987
#> 46  046  0.325227474  1.28990506
#> 47  047  0.067577541  0.28813382
#> 48  048  0.421683474  0.76894727
#> 49  049  0.009598334  0.18377240
#> 50  050 -0.025073035 -0.83195609
#> 51  051  0.302123929 -1.06891040
#> 52  052  0.699079517  0.47224228
#> 53  053  0.621581156  1.19173121
#> 54  054  1.205165948  0.44278118
#> 55  055 -1.433827130 -2.06161251
#> 56  056  0.961075461 -1.48843697
#> 57  057 -1.012911587  0.45703133
#> 58  058 -1.682238924 -2.16346939
#> 59  059  0.661437359  1.82855029
#> 60  060 -0.282607892 -0.09384091
#> 61  061  0.422443322  1.64094654
#> 62  062  0.199541997 -1.65721753
#> 63  063 -1.396205129 -0.97181981
#> 64  064 -0.300053342 -1.14425504
#> 65  065  0.898191437  1.36348193
#> 66  066  0.495065536  0.37500849
#> 67  067 -1.343093789 -0.78692700
#> 68  068  1.327278615 -0.27444099
#> 69  069 -0.486120666  0.08234338
#> 70  070 -0.891626737 -0.31804176
#> 71  071  0.183607087 -1.16133897
#> 72  072  1.212335768  0.93216030
#> 73  073  0.729369563 -0.74078409
#> 74  074  0.405619894 -0.27192800
#> 75  075  1.172534593  0.67594904
#> 76  076  0.205338581 -0.59681030
#> 77  077  1.007388549  0.37806119
#> 78  078 -1.541101448 -1.76287344
#> 79  079 -0.906682596 -0.84312296
#> 80  080  0.614305711  0.81993437
#> 81  081  0.430897207  0.45235200
#> 82  082  0.661274223  0.26455996
#> 83  083  1.444051345  1.01944145
#> 84  084 -0.613180342 -1.26390935
#> 85  085 -0.269723024 -1.02386729
#> 86  086 -1.473138642 -0.19303293
#> 87  087  0.539656980  0.76067505
#> 88  088  0.405229997  0.95981811
#> 89  089  1.199723045  1.51299484
#> 90  090  2.233779362  0.62661688
#> 91  091 -0.865592082 -0.98567547
#> 92  092 -0.023188242 -0.42327583
#> 93  093  0.432953083 -0.26931264
#> 94  094  0.720578860 -0.47926533
#> 95  095  0.897527800  0.59519795
#> 96  096  0.853055368 -0.54664599
#> 97  097  0.014028885 -0.55042842
#> 98  098 -0.570210509 -1.66628773
#> 99  099 -0.835614593  0.08014843
#> 100 100 -0.113523515  1.37458641