Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4744541
#> Y_2 0.4744541 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.3901516  1.3038324
#> 2 -0.1251217  0.7038808
#> 3  1.0141826 -0.1937659
#> 4  0.4107018  0.9762540
#> 5 -0.8559613  0.8113149

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.103223224 -1.234582618
#> 2   002  0.442326477  0.092707222
#> 3   003  0.790632790  0.034061631
#> 4   004  1.344696671  0.483824818
#> 5   005  0.072595497 -0.491441323
#> 6   006 -2.474141873 -1.173347168
#> 7   007 -0.128782684  0.026881897
#> 8   008 -0.294326254 -1.120308107
#> 9   009 -0.862140746 -0.907610268
#> 10  010  0.377169004  0.154622420
#> 11  011  0.034903482 -0.384863782
#> 12  012  1.043212236  0.239191908
#> 13  013  0.797167981 -0.355468587
#> 14  014 -0.499663958  0.008182347
#> 15  015 -0.958806644 -0.602568139
#> 16  016  2.435229186  1.201669840
#> 17  017  0.276290840  0.169896957
#> 18  018  0.532132160  0.101277255
#> 19  019  0.532318256  0.222025515
#> 20  020  0.266846417 -0.804062058
#> 21  021  0.861893206  0.085529932
#> 22  022  0.658542001  0.496824294
#> 23  023  0.217488753 -0.834909424
#> 24  024 -1.600286057 -0.588314066
#> 25  025  1.092784866  0.853954582
#> 26  026  1.027796991  0.364275870
#> 27  027  0.770728033  0.167995314
#> 28  028 -1.020032038 -0.772072511
#> 29  029  0.839660354  1.653417010
#> 30  030 -0.152168992  0.743489465
#> 31  031 -0.816724650  0.586568644
#> 32  032  0.917030196 -0.924370797
#> 33  033  0.656332829  0.508493180
#> 34  034  1.827417337 -0.500074233
#> 35  035  1.042970391  1.335268815
#> 36  036 -1.906798265 -0.652302310
#> 37  037  1.328482087 -0.022712470
#> 38  038 -0.974255536 -0.451490641
#> 39  039 -1.047511831  1.082027084
#> 40  040 -0.090893807 -0.196325219
#> 41  041 -0.949424250  0.166631382
#> 42  042  0.811365849  0.138856237
#> 43  043  1.164439457  0.703834684
#> 44  044  0.415915608  0.952968644
#> 45  045 -1.068199519  1.473264246
#> 46  046  0.917876103  1.218701796
#> 47  047  0.876007938  0.556648353
#> 48  048  0.524277585 -1.584277955
#> 49  049 -0.373433916  0.624285818
#> 50  050 -0.094194660 -0.265067787
#> 51  051 -1.273368098 -1.332582462
#> 52  052  0.119656000 -1.051834639
#> 53  053  0.608101798 -0.068960312
#> 54  054  0.049874269  1.680715617
#> 55  055 -0.346968190 -1.033946378
#> 56  056  0.267632340 -0.023877277
#> 57  057  0.506305639  0.370919999
#> 58  058 -0.180315250  0.582605349
#> 59  059 -1.071578879 -0.401456842
#> 60  060  0.288466944 -0.528643743
#> 61  061 -0.037004435 -0.791614685
#> 62  062  1.579251515 -0.393756509
#> 63  063  0.933688053  0.415904985
#> 64  064 -0.386206200 -1.165993904
#> 65  065 -0.145382692 -0.067064914
#> 66  066  1.510135363  0.649592703
#> 67  067  0.661756099  0.052188388
#> 68  068 -0.437720520  0.319847714
#> 69  069  1.874791278 -0.234478173
#> 70  070  0.752515030 -0.426936096
#> 71  071  0.457037385  0.477483579
#> 72  072  0.001389238 -0.130049128
#> 73  073  0.758701587  1.538186575
#> 74  074 -0.667374685  0.076236349
#> 75  075  0.405230017 -0.215320781
#> 76  076 -0.625689488 -1.989278033
#> 77  077 -0.668433627 -2.011551383
#> 78  078 -0.518816252 -0.144152843
#> 79  079  1.330103915  1.288364607
#> 80  080  2.087094345  1.850534068
#> 81  081 -1.239708956 -1.157346848
#> 82  082  0.965948402  0.864035757
#> 83  083  1.026016191 -0.670592111
#> 84  084 -1.474942558 -0.567225032
#> 85  085  0.425478987 -0.519360298
#> 86  086 -0.335615709  0.709586854
#> 87  087 -0.797914494 -0.567199508
#> 88  088 -0.501341031 -0.137550784
#> 89  089 -0.210207711 -1.545306506
#> 90  090 -0.002294249 -0.134703201
#> 91  091 -0.485414100 -1.251494793
#> 92  092 -0.073182076  0.876922802
#> 93  093  0.439918169  0.729026559
#> 94  094 -0.412564341  0.105899278
#> 95  095  0.547530775  1.666815958
#> 96  096 -0.773628125 -0.824821662
#> 97  097  0.420250693  0.594618769
#> 98  098  0.591613296  1.394824283
#> 99  099  1.070190217  1.775614800
#> 100 100  0.619547678  0.653833797

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -0.490097202  1.435044508
#> 2   002  1.652597132  1.113586953
#> 3   003 -0.590936238  0.514389892
#> 4   004 -0.337814883  0.882560135
#> 5   005  2.455494076  0.459814850
#> 6   006 -0.527781544  0.701970954
#> 7   007 -0.154494201 -0.554650879
#> 8   008  0.173264502 -0.740358918
#> 9   009  0.812965961 -0.328609838
#> 10  010 -1.308922229 -0.974305434
#> 11  011 -0.897318506 -0.931420409
#> 12  012  0.260298950  0.178453928
#> 13  013  1.434730115  0.889846697
#> 14  014  0.894630493  0.552938654
#> 15  015  0.630107137  0.268933644
#> 16  016  1.438912397  0.720737499
#> 17  017  0.662943676 -0.001844241
#> 18  018  0.256646978  1.156620582
#> 19  019  0.104066093 -0.279182762
#> 20  020  0.611947136  0.840197609
#> 21  021 -0.717066376  0.549729694
#> 22  022 -0.156486681 -0.217122442
#> 23  023 -0.162497709 -0.144659899
#> 24  024 -0.482213388 -1.227185596
#> 25  025 -0.859575130 -2.018701104
#> 26  026 -0.877249117 -1.065127154
#> 27  027  0.010077663  0.604334771
#> 28  028 -1.162323659 -0.396653567
#> 29  029 -0.126950825  0.240279296
#> 30  030  0.864974337  1.353634912
#> 31  031 -0.654697155 -0.711824086
#> 32  032  1.006422556  0.055978569
#> 33  033 -0.697497916 -0.256197334
#> 34  034 -0.506152458 -1.190139131
#> 35  035  0.906395023 -1.060115782
#> 36  036 -2.252992693 -1.499685779
#> 37  037 -1.192520821 -0.513201905
#> 38  038 -0.254827297  0.246977755
#> 39  039 -1.525755896 -1.809240339
#> 40  040  0.637112737  0.193393385
#> 41  041  0.693675207 -0.074580366
#> 42  042  0.615265240  0.061719703
#> 43  043  0.004336284 -0.350998012
#> 44  044  1.601408906  1.564594649
#> 45  045  0.640415532  0.157129417
#> 46  046 -1.989621545 -0.945156197
#> 47  047 -0.363638007  0.535333459
#> 48  048  0.215527154  1.393180330
#> 49  049  0.372326053  0.171545192
#> 50  050 -1.036303299 -0.274440411
#> 51  051  1.004938475  1.976680927
#> 52  052  0.866504058  0.333874789
#> 53  053 -0.715886001 -0.242321936
#> 54  054 -0.997233671  0.147171456
#> 55  055 -0.421876565  1.300559183
#> 56  056 -1.786799687 -1.149264403
#> 57  057 -0.214885536  0.950121161
#> 58  058 -0.576926107  0.794473419
#> 59  059 -0.317415829  1.067967639
#> 60  060  0.174459333 -1.119734102
#> 61  061 -0.311867208  0.859505122
#> 62  062 -0.165815818  0.594016682
#> 63  063  0.328929080  0.380538137
#> 64  064  0.006662617 -0.653846115
#> 65  065  0.012087289 -0.425766351
#> 66  066  1.592975679  0.644249975
#> 67  067  1.545447664  0.189584586
#> 68  068 -1.401190597 -0.206758096
#> 69  069  0.738654790  0.214767964
#> 70  070  0.180335637 -0.646273934
#> 71  071 -1.345594272  0.438153667
#> 72  072  0.038544945 -1.370176895
#> 73  073  1.397205614  0.749040880
#> 74  074  1.292511259 -1.639796884
#> 75  075 -0.232176332  0.019742559
#> 76  076  1.084674104  0.106464512
#> 77  077  0.011592246  1.401399788
#> 78  078 -1.899514791 -0.572852474
#> 79  079 -1.430461910 -1.016983038
#> 80  080  0.220640624  0.965674745
#> 81  081 -0.623520794 -1.786798537
#> 82  082  0.696496881  1.686070748
#> 83  083  0.267673766  1.451777680
#> 84  084 -0.488048814 -0.222610305
#> 85  085 -0.717479863 -1.115232930
#> 86  086 -0.127296785 -0.117248072
#> 87  087 -0.382663768 -0.131241502
#> 88  088 -0.372775212  0.254774038
#> 89  089  0.489349361  1.510429590
#> 90  090 -2.555858792 -1.092659428
#> 91  091  1.254765089  0.942880892
#> 92  092  0.144949736 -0.758187441
#> 93  093  0.609818759  0.717181258
#> 94  094  1.176190494  0.043481093
#> 95  095 -0.657633886 -0.730722700
#> 96  096 -0.091786857  2.013576252
#> 97  097 -1.122995521  0.895119135
#> 98  098  2.215540310  2.296861837
#> 99  099  0.088005580 -0.567272793
#> 100 100 -1.277123295 -0.987323525