Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5469335
#> Y_2 0.5469335 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5227545  0.5228770
#> 2  2.8850571  1.5679917
#> 3  0.1132045 -0.2204946
#> 4  0.8071490 -0.7807221
#> 5 -0.6894401  0.1444927

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.338758876  1.441765191
#> 2   002  0.254627718  1.773333311
#> 3   003  0.332628033 -0.227962433
#> 4   004 -0.105128280  2.381958325
#> 5   005 -1.319676142  0.351394358
#> 6   006  0.994249103  0.551877644
#> 7   007  1.006457057 -0.531290324
#> 8   008 -0.499721353 -1.158679828
#> 9   009 -1.581417181 -0.153005508
#> 10  010 -2.437239477 -0.171633297
#> 11  011 -0.867872714 -0.475487951
#> 12  012 -1.834965940 -0.930489893
#> 13  013  0.032168980  0.410616452
#> 14  014  0.436554133 -0.390255831
#> 15  015  0.850315759  0.696654331
#> 16  016 -0.584458102 -0.066051810
#> 17  017  1.639939471 -0.291004899
#> 18  018  0.325809048  0.207213375
#> 19  019 -0.862893036 -0.628523557
#> 20  020 -0.887101989 -0.221396669
#> 21  021  0.585637181  1.299161226
#> 22  022 -1.473910703 -1.037354119
#> 23  023  0.545457889  0.290400864
#> 24  024 -0.175077477 -0.622501954
#> 25  025  0.556894290 -1.085647785
#> 26  026 -0.544991431 -0.703194757
#> 27  027 -0.871425617 -1.117263929
#> 28  028  0.137413834 -0.214220271
#> 29  029 -1.122847712  0.810048700
#> 30  030  0.279388900 -1.556787752
#> 31  031 -2.173373129 -0.774112045
#> 32  032 -0.680526232 -0.252353460
#> 33  033 -0.451278337 -0.653490290
#> 34  034  0.696846373  0.774705462
#> 35  035  1.159164867  0.903711790
#> 36  036 -1.338023391 -0.706715650
#> 37  037  0.007717293 -0.408515984
#> 38  038 -1.306644290 -1.596008059
#> 39  039 -0.426540449 -0.541597447
#> 40  040 -2.028637176 -0.296406825
#> 41  041  0.390180181  1.271380565
#> 42  042  1.686314880  2.233881978
#> 43  043 -0.889667327  0.280176768
#> 44  044 -0.498882396  0.171971129
#> 45  045  1.330628976 -1.003406301
#> 46  046 -1.093754886 -0.687949855
#> 47  047 -1.068950876  0.909012809
#> 48  048  0.805868328 -0.824052325
#> 49  049  0.419547187  1.446953087
#> 50  050  0.299321387 -0.113263624
#> 51  051 -2.119392343  0.047514806
#> 52  052 -0.705527658 -0.176471867
#> 53  053  0.536629960  1.722139548
#> 54  054 -0.412145275  1.339654311
#> 55  055  0.217628134 -0.373845175
#> 56  056 -1.081161803 -1.114316524
#> 57  057 -0.311785902  0.363520095
#> 58  058 -1.246946439 -2.290107848
#> 59  059 -0.898915166 -1.040691379
#> 60  060  0.471344802  0.009516934
#> 61  061 -1.175350717 -1.249684130
#> 62  062 -0.344546981 -0.298749733
#> 63  063 -0.642283527 -0.534502514
#> 64  064  0.076919248  0.841983256
#> 65  065  0.498966664 -0.556222239
#> 66  066  0.142681837 -1.635715545
#> 67  067  0.158169949  0.110186937
#> 68  068 -0.306808417 -0.231252927
#> 69  069  0.419234548  1.070189068
#> 70  070  1.018561620  0.010492113
#> 71  071 -0.904219045 -0.350544245
#> 72  072 -0.961889539 -0.067387672
#> 73  073 -0.653124874 -0.931039575
#> 74  074  1.849129556  1.052819773
#> 75  075  0.850641112  0.368062470
#> 76  076 -0.102090582  0.809963235
#> 77  077  0.837027904 -1.903793934
#> 78  078  0.652718414 -0.038248302
#> 79  079 -1.149355321 -1.194057567
#> 80  080 -1.108433049 -1.299776309
#> 81  081  0.366001683 -0.162746309
#> 82  082 -0.985426365 -1.005344127
#> 83  083 -0.174979654  0.393587480
#> 84  084  2.320000751  0.842041418
#> 85  085  1.475464055  1.962268861
#> 86  086 -0.447735081 -0.381078710
#> 87  087  0.123095956 -0.323056870
#> 88  088  1.599929680 -1.036611052
#> 89  089  0.090176677  0.540221027
#> 90  090  0.085517646 -1.060670953
#> 91  091  0.238603197  1.146126030
#> 92  092 -1.112888471  0.289932830
#> 93  093 -0.213945636 -1.888493933
#> 94  094 -0.007610287  0.573506747
#> 95  095  0.663283341 -0.193042822
#> 96  096 -0.839631680 -1.704186589
#> 97  097  0.146677911 -0.420731203
#> 98  098  0.141425287 -0.306509875
#> 99  099  1.432366493  1.740109620
#> 100 100  0.056374325  0.538807241

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  1.299399993  0.275600080
#> 2   002 -1.461651450  0.292332810
#> 3   003  1.423926218  0.884765238
#> 4   004  1.005053646  0.938859803
#> 5   005 -2.991059120 -1.535639113
#> 6   006  0.853766670  0.545450182
#> 7   007 -0.232328800 -1.203542503
#> 8   008 -0.171434126  0.403663209
#> 9   009  0.609804624  0.336388601
#> 10  010  1.119298092  1.310465492
#> 11  011 -0.382717167  0.606075062
#> 12  012 -0.333098879  0.217493809
#> 13  013  0.168525690 -0.622789545
#> 14  014  1.147047692 -0.055650738
#> 15  015  0.194419180 -0.113350218
#> 16  016  0.414610777  1.463498845
#> 17  017  0.560528388  0.505555672
#> 18  018  0.711952471  0.588487640
#> 19  019  0.412269998 -1.623329395
#> 20  020  0.832309770 -0.466720572
#> 21  021 -0.036966232  0.379222839
#> 22  022 -0.726425969  1.566076470
#> 23  023  0.012885852  1.022918481
#> 24  024  0.236290642  1.375479447
#> 25  025 -0.195304690 -0.677514416
#> 26  026 -0.519063706 -0.489943286
#> 27  027  1.055349664 -2.347882373
#> 28  028  0.776957241  2.263844242
#> 29  029 -1.937360207 -1.303129470
#> 30  030  0.215532704 -0.155227409
#> 31  031 -0.853742596  0.368620211
#> 32  032 -1.036809365 -1.401158201
#> 33  033 -2.558290687 -1.650122201
#> 34  034 -0.310281865  0.325256832
#> 35  035  1.550102532  0.928725771
#> 36  036 -0.157506820  1.026689462
#> 37  037  0.095083206  0.595568447
#> 38  038  0.196457231  1.438855785
#> 39  039  0.154134448 -0.685905573
#> 40  040 -0.500543905 -0.746649394
#> 41  041 -0.254723769 -0.231909530
#> 42  042 -0.817217231  0.616399241
#> 43  043 -1.212751839 -1.543375567
#> 44  044 -0.258356987  0.053736472
#> 45  045 -2.174434959 -0.531925382
#> 46  046 -1.171248391 -1.039490037
#> 47  047 -1.177712698  0.495441138
#> 48  048 -0.121930652  0.156809740
#> 49  049  0.177609271  0.077456039
#> 50  050 -0.247202133  0.031322787
#> 51  051 -0.119348925  0.207896683
#> 52  052  0.045433213  0.177070638
#> 53  053  0.202837977  0.364194972
#> 54  054 -0.878092492 -0.448108891
#> 55  055 -0.118495322  0.394607004
#> 56  056 -0.264118926 -0.262669572
#> 57  057  0.849862298  0.755965641
#> 58  058  0.454871284  1.147590055
#> 59  059  1.845990625 -0.276920253
#> 60  060 -0.102791869  0.108331800
#> 61  061 -0.006860174  0.113637895
#> 62  062 -0.805617691 -1.429172828
#> 63  063 -1.652156240 -1.446871840
#> 64  064  1.338661476 -0.047154935
#> 65  065  1.480877686  0.982904488
#> 66  066  0.611158927 -0.124515684
#> 67  067 -0.973902084 -0.449807084
#> 68  068 -0.912876834 -0.938862997
#> 69  069 -0.693708759 -2.090470399
#> 70  070 -0.070990829 -0.217798681
#> 71  071 -0.669543242 -1.009117046
#> 72  072  0.136269940  1.097741100
#> 73  073 -0.389232868 -1.249413985
#> 74  074  0.033070821  0.545886718
#> 75  075 -0.059636509  0.704487209
#> 76  076 -0.108576686 -0.609745445
#> 77  077  2.263604750  2.768607353
#> 78  078  0.328820588 -0.316437564
#> 79  079  0.289288832  0.385051608
#> 80  080  0.072913461  0.102708365
#> 81  081  0.189406947  1.334000082
#> 82  082  1.446228230  0.725588673
#> 83  083 -1.819756055 -1.005987531
#> 84  084 -0.288142819 -0.715240002
#> 85  085  0.763906746 -0.009099103
#> 86  086 -0.547615197 -0.633910250
#> 87  087 -1.226082181  0.793296990
#> 88  088  0.679188721  0.164612437
#> 89  089 -0.786393345  0.080881555
#> 90  090  0.722574600  1.652475374
#> 91  091 -1.450878163  0.538577727
#> 92  092 -0.807114172 -2.007534826
#> 93  093  1.982863009  1.105303460
#> 94  094 -1.121274848 -1.483287693
#> 95  095 -0.905818353 -1.505994976
#> 96  096  1.407667819  0.679253991
#> 97  097  1.262251777 -0.170245864
#> 98  098 -0.073968908  1.290383359
#> 99  099 -1.271837110 -0.352418174
#> 100 100  0.362241660 -0.459601809