Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4667366
#> Y_2 0.4667366 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4784861 -0.5751678
#> 2  0.9235585 -1.3951639
#> 3 -0.2491991  1.2009527
#> 4  0.2538765  0.6469841
#> 5 -0.1187819  0.2733889

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.01735195 -2.228920170
#> 2   002  0.65449729  0.757367076
#> 3   003  0.75056701  0.609422722
#> 4   004  0.99403447  1.623840709
#> 5   005 -0.21222206  0.386488615
#> 6   006  1.91731133  1.773375294
#> 7   007  0.14265770  0.381064643
#> 8   008 -1.11837853 -0.495577287
#> 9   009  1.59603833  0.014005314
#> 10  010 -1.39523876 -1.674125151
#> 11  011  1.08099326  1.145933479
#> 12  012 -0.82068161 -1.954828323
#> 13  013 -0.79008936  1.457929038
#> 14  014  1.00284639  0.727230459
#> 15  015 -0.34794707 -1.087018561
#> 16  016 -0.21133894  0.916758864
#> 17  017 -0.79296446  0.400350115
#> 18  018  0.04994775  0.614069353
#> 19  019  0.54635837 -1.510152981
#> 20  020 -0.44317913 -0.894744619
#> 21  021 -0.93850119 -1.940591150
#> 22  022  0.37690769 -0.444352665
#> 23  023 -0.35648483 -0.018970388
#> 24  024  0.76471470  1.182635046
#> 25  025  0.28099905  0.266411684
#> 26  026  0.90485819 -1.904686432
#> 27  027  2.26805708  1.544170457
#> 28  028 -0.51056690 -0.433118263
#> 29  029  0.81274931 -0.123802639
#> 30  030  0.36160520  0.434524854
#> 31  031 -0.86717803 -0.041330399
#> 32  032  0.14531664  0.531666882
#> 33  033  3.24600854  1.824373177
#> 34  034  2.65485997  1.163528895
#> 35  035 -1.29027682 -2.162892319
#> 36  036 -1.02818515 -0.615697635
#> 37  037 -0.12536373  0.372846791
#> 38  038  0.97728146  0.049860134
#> 39  039 -1.22423280 -1.524425312
#> 40  040  0.45333399 -1.494367705
#> 41  041  0.49314795  0.561782149
#> 42  042 -0.18129972  0.526095117
#> 43  043 -1.71152467 -1.757551286
#> 44  044 -0.73185463  0.379385393
#> 45  045 -2.04732937 -1.451908339
#> 46  046  2.33733711  0.804704327
#> 47  047 -0.17371883 -0.003785646
#> 48  048 -1.80549115 -1.336752821
#> 49  049 -0.64738337 -1.669487045
#> 50  050 -0.82752247 -0.507204343
#> 51  051 -0.53684289  1.161673171
#> 52  052 -2.29014390 -1.933434599
#> 53  053  0.27167283 -0.298259607
#> 54  054 -0.72354690 -1.978348392
#> 55  055  1.19852651  2.409643000
#> 56  056  1.99881233  0.914927819
#> 57  057 -0.49254968  0.334400937
#> 58  058 -0.64539682 -1.268869628
#> 59  059  1.45801320  1.200689311
#> 60  060 -0.99054388 -0.338756599
#> 61  061  0.62262777  0.166531255
#> 62  062  0.95255755  0.552112873
#> 63  063 -0.85774291  0.693509676
#> 64  064 -1.07437715 -0.414389074
#> 65  065  1.46090666 -0.720418513
#> 66  066  1.02390747  0.165493651
#> 67  067 -0.50013145  1.293914791
#> 68  068 -0.49802913 -0.667272488
#> 69  069  0.78377556  0.322136120
#> 70  070  0.32141306 -0.395231828
#> 71  071 -1.11861633 -1.042933077
#> 72  072 -0.47348451  0.329110906
#> 73  073  0.21586655  0.165906520
#> 74  074  1.29232070 -1.174755375
#> 75  075  0.01858218 -0.839187410
#> 76  076  1.30207271  0.578997992
#> 77  077 -0.58864913  0.416461893
#> 78  078 -1.11000251 -1.645540264
#> 79  079 -0.18038752 -0.174019759
#> 80  080  0.52064160 -1.175156939
#> 81  081 -1.73536842  0.090503962
#> 82  082  0.70577122  0.999326148
#> 83  083 -1.21888353 -1.237723979
#> 84  084  0.11544923  1.572762471
#> 85  085 -0.34940913 -0.448925177
#> 86  086  1.10712464  0.413920484
#> 87  087 -0.32035917  0.076296880
#> 88  088  0.60549008  1.176929679
#> 89  089  0.03245041 -0.578880064
#> 90  090 -0.70628460  0.249536418
#> 91  091  0.94932572 -0.612966519
#> 92  092  1.29494447  1.482924942
#> 93  093  0.53387031  0.350590043
#> 94  094  0.49187812 -1.970603905
#> 95  095  0.45588107  0.677305430
#> 96  096  1.05991874  1.218490191
#> 97  097  1.25717313  1.445834367
#> 98  098  0.55209918  0.941406017
#> 99  099  1.05328193  0.003183332
#> 100 100  1.22647982  0.374150323

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.59394736 -0.64957249
#> 2   002  0.28280036  0.07918229
#> 3   003  0.11115363  0.10438738
#> 4   004 -0.31977722 -0.77297028
#> 5   005  0.49326781  0.22502173
#> 6   006 -0.79980773 -0.34083338
#> 7   007  1.08258150  2.07761081
#> 8   008 -0.03255977 -0.56026250
#> 9   009 -3.05220392 -0.98304319
#> 10  010 -0.07709784  1.13944283
#> 11  011  0.88053429 -0.14173873
#> 12  012  0.19140840 -0.58379685
#> 13  013  0.36664534  1.34044382
#> 14  014 -0.32829706 -0.81515860
#> 15  015  0.80168023  1.31459321
#> 16  016  0.56094434 -2.18887440
#> 17  017 -0.43448574  1.30679397
#> 18  018 -0.84808917 -0.58082281
#> 19  019 -0.28993209  0.48240883
#> 20  020 -1.06244878 -0.82054022
#> 21  021  1.30468721  0.59676394
#> 22  022  0.36097244 -0.27698062
#> 23  023 -1.45892878 -0.10828466
#> 24  024  1.10738063  1.27605970
#> 25  025 -0.58173249  0.25052926
#> 26  026  0.66059807  1.29291792
#> 27  027  0.15474556 -0.80597475
#> 28  028 -0.87562652 -1.77695466
#> 29  029  0.68734808 -0.52073851
#> 30  030 -1.80412573 -1.33621090
#> 31  031  1.79231968  1.09252644
#> 32  032  1.83762392  0.20619166
#> 33  033 -0.81330088 -1.06931919
#> 34  034  0.53273639 -1.99385254
#> 35  035 -0.88796072  0.43066686
#> 36  036 -0.72620124  0.54186315
#> 37  037  0.91434518  0.64241191
#> 38  038 -0.05426563  0.04080048
#> 39  039  0.08284616 -0.29093196
#> 40  040  0.38247175  1.80892915
#> 41  041 -1.64708967 -0.60500461
#> 42  042 -0.56613426 -0.29768339
#> 43  043 -0.41773890 -0.61759843
#> 44  044  1.05677314 -1.03219778
#> 45  045 -0.61877812  0.62813337
#> 46  046  0.65738023 -0.30798546
#> 47  047  0.45968710  0.98190804
#> 48  048 -0.86653256 -2.44038007
#> 49  049 -2.31392185 -1.69301722
#> 50  050 -0.46561936  0.91531448
#> 51  051 -1.21271475 -1.38631910
#> 52  052 -0.54886659 -0.69873925
#> 53  053 -0.86408468 -1.21705516
#> 54  054 -0.13850302 -1.11873247
#> 55  055  1.33324833 -0.08702274
#> 56  056  1.68169168  1.62385879
#> 57  057  0.01569597  0.78390767
#> 58  058 -1.69627362 -1.21988160
#> 59  059  1.43444307  1.48467595
#> 60  060 -0.78895798 -0.71795342
#> 61  061 -1.50365025  1.96696799
#> 62  062 -0.97393875 -2.54545224
#> 63  063  1.61858717  0.25486238
#> 64  064 -0.53829655 -0.00514122
#> 65  065  1.42238188  0.51064593
#> 66  066  0.08237684 -0.41372895
#> 67  067 -0.85051640 -0.60034634
#> 68  068 -1.66099137 -2.40993071
#> 69  069 -0.56370621 -0.88722584
#> 70  070 -0.71058102 -1.09241793
#> 71  071 -0.85217031 -0.71319215
#> 72  072  1.28290637  1.57547040
#> 73  073  0.22422823  0.36331382
#> 74  074 -0.52873672 -0.56586952
#> 75  075  1.13442839  0.55639472
#> 76  076  0.81854067 -0.33522613
#> 77  077 -1.02333535 -1.55305502
#> 78  078 -0.18257254  0.23942766
#> 79  079 -0.10372204 -1.04221830
#> 80  080  0.39669468 -0.10920511
#> 81  081  0.30280835  0.15878619
#> 82  082 -1.15751583 -0.22384364
#> 83  083 -0.46208624  0.59676110
#> 84  084  0.27202957 -0.91390811
#> 85  085 -0.56041884 -1.45190112
#> 86  086  0.22637487 -0.69059020
#> 87  087  0.83495581  0.09992402
#> 88  088 -0.38604048  1.23971437
#> 89  089  0.08982388  0.55853823
#> 90  090 -1.68086586 -0.08318870
#> 91  091 -0.56870902 -0.89521742
#> 92  092 -0.71949562  0.02664366
#> 93  093 -0.09847915 -1.45612298
#> 94  094  0.82700851  1.10824415
#> 95  095  1.25655409  1.31392912
#> 96  096  1.03151174  0.83290251
#> 97  097 -1.42731600 -1.01303923
#> 98  098 -0.94740678 -0.72553639
#> 99  099  0.89911820  0.64839359
#> 100 100 -0.75347722 -0.38297803