Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4735714
#> Y_2 0.4735714 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.30336176 -0.4814408
#> 2  0.05709041  0.9003596
#> 3  0.52376414  1.5996923
#> 4  1.05551052  0.3169631
#> 5  0.49936945  1.5252826

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.86433849  1.39747003
#> 2   002  0.33928186 -0.84472684
#> 3   003 -0.47108354 -0.26412053
#> 4   004 -0.42125639 -1.23520708
#> 5   005  0.88835145  0.43192369
#> 6   006 -1.53142044 -1.45113389
#> 7   007  0.86912330 -0.25483353
#> 8   008 -1.47625502  0.52220213
#> 9   009  0.01848131  1.69140949
#> 10  010 -0.49067306  0.22534669
#> 11  011  0.15064523 -0.79893051
#> 12  012  0.78038930  1.82806586
#> 13  013 -0.75023397  1.59449288
#> 14  014 -0.90781786 -0.83356620
#> 15  015 -0.81805923  0.37294630
#> 16  016 -0.08495341 -0.65864703
#> 17  017 -0.60057379 -1.11067993
#> 18  018 -0.77385714 -0.51535920
#> 19  019  0.32596361  1.73054842
#> 20  020 -0.13196569 -1.23891862
#> 21  021 -0.16157676  0.92542279
#> 22  022  0.83602471  1.55092076
#> 23  023  1.39888330  1.44635372
#> 24  024 -1.34371562 -1.80813141
#> 25  025  0.58888378  0.39160255
#> 26  026 -0.50801886 -0.52337610
#> 27  027  0.02693500  0.47853472
#> 28  028 -1.41636149 -0.84111536
#> 29  029  1.03884427  1.06157244
#> 30  030 -0.19282107  0.03775145
#> 31  031 -0.17244910  0.31969809
#> 32  032  1.41342891  0.27945436
#> 33  033  0.55204366 -0.71970149
#> 34  034  0.62848677 -0.53974628
#> 35  035  0.60785982  0.45318535
#> 36  036 -1.27208241 -0.22365921
#> 37  037  0.92769483 -1.23610647
#> 38  038  0.64927682  0.43952879
#> 39  039 -0.39562658 -1.05523128
#> 40  040 -0.22972588  0.36400927
#> 41  041  0.39075469  1.97214807
#> 42  042 -0.79514128 -2.71171310
#> 43  043  0.79928705  2.36186458
#> 44  044 -1.79658888 -1.91076890
#> 45  045 -1.27030371 -1.29835364
#> 46  046  0.72668100 -0.50430483
#> 47  047 -0.81151730 -1.47161344
#> 48  048 -0.52504592  1.24082083
#> 49  049 -0.63827956  0.42614485
#> 50  050 -2.29371116 -0.58020824
#> 51  051  0.63763451  0.03126228
#> 52  052  1.65525206  1.58244847
#> 53  053 -1.48549607 -2.30988042
#> 54  054 -2.42137044 -0.64671675
#> 55  055 -0.64736145 -0.13106793
#> 56  056  0.30220095  0.31001370
#> 57  057  0.44513573 -0.19072046
#> 58  058 -1.80250809 -2.30334105
#> 59  059  1.09931007 -0.05465043
#> 60  060 -1.10267558 -2.11757885
#> 61  061 -0.87415797  0.28109893
#> 62  062 -2.41635414 -1.92943679
#> 63  063 -0.39259339  0.30433416
#> 64  064  0.90071681  0.35205022
#> 65  065 -1.19889172 -1.91698936
#> 66  066 -0.99775036 -1.65928875
#> 67  067  0.10846409 -0.36438810
#> 68  068 -0.95531517  0.16382810
#> 69  069  1.29386890  1.95646717
#> 70  070 -0.09116565  0.06137174
#> 71  071 -1.03109065 -1.49716863
#> 72  072 -1.48163458  1.53500087
#> 73  073 -0.53269299 -0.38501631
#> 74  074  0.17881772  0.49951888
#> 75  075 -0.67324048  0.46301005
#> 76  076 -0.24709281  1.18303427
#> 77  077  0.45087033  0.44766382
#> 78  078 -1.04473426 -0.46366887
#> 79  079 -1.59910147 -0.98042206
#> 80  080  1.44067568  1.58907830
#> 81  081  1.20433741  1.18624171
#> 82  082  0.68866015  0.23604787
#> 83  083 -0.41167049 -0.79940448
#> 84  084  1.04911666  0.21967622
#> 85  085 -0.76910040 -2.47267214
#> 86  086 -0.70242383 -0.81110808
#> 87  087  1.88917299 -0.11277757
#> 88  088 -1.37503187  0.52916454
#> 89  089 -1.59529522 -1.25775467
#> 90  090  0.06181268 -0.87228815
#> 91  091  0.92092514  1.15385886
#> 92  092 -2.26731258 -0.54015808
#> 93  093 -1.48356204 -0.95482770
#> 94  094  1.23391385  0.48595756
#> 95  095 -0.08208576  0.07517416
#> 96  096  0.44991971 -0.52457263
#> 97  097 -1.10586410 -0.04106734
#> 98  098 -0.96621878  0.21375349
#> 99  099 -2.05742891 -0.81573255
#> 100 100  0.33019172 -0.29677278

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.88509504  0.746594566
#> 2   002 -1.44555946 -1.138142186
#> 3   003  0.21203913 -0.008271518
#> 4   004  0.11613540  0.534212317
#> 5   005  0.64366924  0.558556169
#> 6   006  0.04244334 -0.485436885
#> 7   007  1.43138178  0.823760007
#> 8   008  1.02812527  2.031137543
#> 9   009 -0.40237018  1.307240456
#> 10  010  2.50198571  1.735331800
#> 11  011 -2.35288811 -1.525523729
#> 12  012  1.33642874  1.566885666
#> 13  013 -0.07654775 -0.503306575
#> 14  014  0.10318002  0.694021091
#> 15  015 -1.19900607 -0.384174332
#> 16  016 -1.13846700  0.312684988
#> 17  017  0.91187056 -0.294823789
#> 18  018 -0.15967816  0.279911318
#> 19  019 -2.29734661 -0.327600691
#> 20  020  2.27550893  0.631311108
#> 21  021  0.68293412  0.266341977
#> 22  022 -1.52877505 -0.333190482
#> 23  023  0.04300774 -0.058747590
#> 24  024 -1.71303249 -1.514091367
#> 25  025 -1.21813581 -1.149139799
#> 26  026 -1.41679672 -0.668781552
#> 27  027 -1.61467257  0.514395815
#> 28  028  0.30527656  0.723588812
#> 29  029 -0.68351847  1.251547244
#> 30  030  1.48816267 -0.130171701
#> 31  031 -0.70374924 -0.538180474
#> 32  032 -0.69592118 -0.699818655
#> 33  033  0.45667279  1.204702574
#> 34  034  1.42935461  0.741388748
#> 35  035  0.67571466  0.012618433
#> 36  036 -0.27839178  1.816818788
#> 37  037  0.07654816 -1.859509659
#> 38  038  0.82716036 -0.491894384
#> 39  039 -1.71414964 -0.693894836
#> 40  040  1.25773878  2.135424600
#> 41  041 -1.42064556 -2.332560903
#> 42  042  1.99987289  0.165819991
#> 43  043  0.03067913  1.607760604
#> 44  044 -1.79767296 -2.076222294
#> 45  045 -1.38554103  0.743714788
#> 46  046 -2.13347776 -1.582993864
#> 47  047 -0.31828717  0.945889747
#> 48  048 -1.11175813 -0.433681351
#> 49  049  1.26221784  0.281495243
#> 50  050  1.58763038  2.223028980
#> 51  051 -1.86084947  0.179499468
#> 52  052  0.74481094  1.192897399
#> 53  053  0.48932532  0.793231975
#> 54  054  0.62472142  0.369021547
#> 55  055 -1.34723899 -2.202712141
#> 56  056 -0.46995856  1.732742293
#> 57  057 -0.35905947 -0.446862058
#> 58  058  0.99235432 -0.345438782
#> 59  059  0.67280948  1.374692155
#> 60  060 -1.38325012 -1.021647031
#> 61  061  1.38006448 -0.244495241
#> 62  062 -0.32017657 -0.810346505
#> 63  063 -0.67124892  0.714833826
#> 64  064 -1.02077439 -0.156818029
#> 65  065  1.12864450 -0.078839553
#> 66  066  0.25642993  0.688683584
#> 67  067  0.09420306  1.420680276
#> 68  068  1.14496897 -0.563827864
#> 69  069  0.10841108  0.216675835
#> 70  070  0.57418712  1.280962657
#> 71  071  0.03811416 -0.196953669
#> 72  072 -1.91560690 -0.193460402
#> 73  073  0.66976728  0.280964704
#> 74  074 -0.16142373  0.702169962
#> 75  075  0.09821446 -0.034925154
#> 76  076 -0.31202749 -0.509919153
#> 77  077  1.15650764  1.416707469
#> 78  078  0.79821672  0.113143103
#> 79  079 -1.42225976 -0.140577454
#> 80  080  0.60963366 -0.255154918
#> 81  081  1.35245795  0.818939389
#> 82  082  0.46881887 -0.883375286
#> 83  083 -0.45490294  1.167931178
#> 84  084  0.30827546  0.151951855
#> 85  085  0.90773619 -0.188750621
#> 86  086 -0.49091938 -0.194016303
#> 87  087  1.15821215 -1.289923816
#> 88  088 -0.63664074  0.258832159
#> 89  089 -1.47734142  0.330435513
#> 90  090  0.60636235  0.121167760
#> 91  091  0.29761366 -0.413120357
#> 92  092  2.49632601  1.082307828
#> 93  093 -3.10113090 -1.951945104
#> 94  094  0.37932866 -0.006260186
#> 95  095 -0.58260999  1.153660977
#> 96  096  1.56336620  1.140549872
#> 97  097  0.56513931  0.864412527
#> 98  098 -0.17692693  0.670407341
#> 99  099  0.08497074 -0.716977101
#> 100 100  0.39252565  0.335094000