Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4683731
#> Y_2 0.4683731 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.7220667  1.2110291
#> 2  0.5107554  0.4290675
#> 3 -0.2425235 -0.6542629
#> 4  0.3504939  0.9183162
#> 5  1.2677586 -0.5712487

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.15510131 -0.421742622
#> 2   002 -2.45271483  0.391581156
#> 3   003 -0.18955441  0.059451849
#> 4   004 -0.31705418 -0.346317268
#> 5   005  1.19739552  1.793653451
#> 6   006 -0.05495538  0.003868166
#> 7   007 -0.42872346 -0.150759903
#> 8   008 -0.44218702  0.055872323
#> 9   009 -0.79208677 -0.976887242
#> 10  010  1.24051778  0.587557286
#> 11  011  0.93972000  0.423083212
#> 12  012  0.21622733  0.223634750
#> 13  013 -0.22404895  0.044417345
#> 14  014  2.49161147  1.415518198
#> 15  015  0.80093880 -1.891782243
#> 16  016  0.05266936  1.323285302
#> 17  017 -0.48918727 -0.089937373
#> 18  018 -0.64747988  0.844412489
#> 19  019  0.40266263 -0.348843380
#> 20  020 -0.81319977 -0.990012845
#> 21  021  1.39319696  1.194655542
#> 22  022  0.00264883  0.702561811
#> 23  023 -0.08826987 -0.082174147
#> 24  024  0.23925375  0.848908755
#> 25  025  0.48881295 -0.066896505
#> 26  026 -0.13640819  0.139272484
#> 27  027  0.38189056  0.329701191
#> 28  028 -0.36210700 -1.476613030
#> 29  029 -0.02204444  1.065253199
#> 30  030  0.13658293  0.069944419
#> 31  031  1.97693696  1.079819742
#> 32  032  0.25517524  0.415548133
#> 33  033  0.41586913  0.182286422
#> 34  034 -0.19275677  0.459024249
#> 35  035 -0.48578496  0.675630428
#> 36  036 -1.07083567 -0.448194438
#> 37  037 -0.01122453 -1.058568021
#> 38  038  0.89803470 -0.218261566
#> 39  039  1.09353075 -0.540330913
#> 40  040  0.25435050  0.185646775
#> 41  041  0.10537489 -0.558276288
#> 42  042 -0.85367026 -0.026289187
#> 43  043  0.90427464  1.245634723
#> 44  044  1.74879978  1.946835204
#> 45  045  0.51840180  0.177612325
#> 46  046 -1.18360447  0.020695300
#> 47  047  1.25295738  2.320030425
#> 48  048 -0.97772664  0.731902410
#> 49  049  1.54712066  2.138289692
#> 50  050  0.55379522  1.083697436
#> 51  051 -0.88945043 -0.476212751
#> 52  052  1.61953318  1.192606682
#> 53  053 -1.02596086  0.364300127
#> 54  054 -0.30806610  0.373671011
#> 55  055 -0.01690931  0.237427444
#> 56  056 -0.51070030 -0.838255552
#> 57  057  0.38810226 -1.423097078
#> 58  058 -0.50690729 -0.560004599
#> 59  059  1.84946019  2.330219389
#> 60  060  0.03006952 -0.088348212
#> 61  061 -0.62742021 -0.120572270
#> 62  062 -1.44508400 -0.999148067
#> 63  063 -1.16315921 -1.507309790
#> 64  064  1.70585845  0.256877024
#> 65  065 -1.06853149 -0.278851027
#> 66  066 -0.19510629  1.033052266
#> 67  067  1.99443265  0.711185767
#> 68  068  0.07049492 -0.017098144
#> 69  069  0.98545929 -0.202281043
#> 70  070  0.93558226 -0.091712895
#> 71  071  0.72033816 -0.345278935
#> 72  072 -0.73960067  0.195681574
#> 73  073 -0.85855547 -0.368392692
#> 74  074 -0.28092377  0.647308228
#> 75  075 -0.09453275 -0.405571812
#> 76  076  0.96453004  1.098367416
#> 77  077 -0.60686777  0.396342529
#> 78  078  0.83553490  0.097525502
#> 79  079  0.54470565  0.797201570
#> 80  080  0.88759436  0.292322150
#> 81  081 -0.39651154  1.047193901
#> 82  082 -0.74392644 -0.612971097
#> 83  083  0.99744717  2.348304592
#> 84  084 -0.55178624 -0.716606721
#> 85  085 -0.71370677 -0.802226725
#> 86  086  0.53954253 -0.411909529
#> 87  087 -0.78159957  0.447866594
#> 88  088 -0.21510744 -2.229372614
#> 89  089  0.14715924  0.806329118
#> 90  090  0.03126814  0.159329253
#> 91  091  1.37680243 -0.354753967
#> 92  092 -0.58730349 -0.923683116
#> 93  093 -2.73601430  0.583514720
#> 94  094  0.09893951 -0.574736182
#> 95  095 -1.42886224 -1.274383587
#> 96  096 -0.54775563 -0.644493584
#> 97  097  1.22926053  0.640566690
#> 98  098  0.05567836 -0.320408308
#> 99  099  0.02783618  0.436335704
#> 100 100 -0.76409505 -0.595447170

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.11635730 -1.59593704
#> 2   002 -0.71296690  0.35617648
#> 3   003  0.95591337  0.04764214
#> 4   004  0.43241575  0.49441678
#> 5   005 -0.69136713 -1.04235690
#> 6   006  1.97397183  2.04801685
#> 7   007  0.20903231  0.46889922
#> 8   008  0.14483353 -0.52867643
#> 9   009  0.04824763 -0.60677814
#> 10  010 -1.64851558 -0.76378497
#> 11  011  1.39031933  0.31562308
#> 12  012 -0.03493743 -0.01084388
#> 13  013  0.64382327  1.10130466
#> 14  014  0.28963785  0.68259955
#> 15  015 -0.94629187 -1.28254763
#> 16  016  0.06855003  0.44032744
#> 17  017  0.90525394  0.39560427
#> 18  018  0.18040251  0.16245794
#> 19  019  2.85513490  2.25749968
#> 20  020  0.44280030  2.29507116
#> 21  021  0.82226407  1.57823742
#> 22  022  0.71962805  1.64969716
#> 23  023 -0.06400606  0.22831693
#> 24  024 -0.31243626  1.08930529
#> 25  025  0.95409919 -0.31117925
#> 26  026  0.64576100 -0.98120439
#> 27  027 -1.75620233 -0.78929688
#> 28  028 -0.26384811  0.33167939
#> 29  029  1.06179535  0.62636713
#> 30  030  0.42590284  1.47135115
#> 31  031 -0.85383766 -0.32106531
#> 32  032 -1.12468798 -0.65825402
#> 33  033 -1.52959600  0.70541989
#> 34  034  0.25852830 -0.11505597
#> 35  035  1.56929422  0.13958564
#> 36  036  0.74562807  1.38704961
#> 37  037  1.43588560  0.40246647
#> 38  038 -0.05570072  0.11621736
#> 39  039  0.15218408 -1.92212458
#> 40  040  0.72586941  1.06645998
#> 41  041  0.24548368  0.93139221
#> 42  042  1.61298046 -1.29527689
#> 43  043  0.37765029 -1.26756120
#> 44  044 -0.48515841 -1.90970887
#> 45  045 -0.38010954 -0.73466199
#> 46  046  0.44022813  0.77116868
#> 47  047 -0.54723371 -0.50539240
#> 48  048  0.12697277 -0.19699640
#> 49  049  0.32759512 -0.01761711
#> 50  050 -0.82721718 -0.99700526
#> 51  051 -0.76666883 -0.93265520
#> 52  052 -1.40013219 -0.08838691
#> 53  053 -0.67668407  0.11286505
#> 54  054 -0.48393645  0.18490683
#> 55  055  0.31949100 -0.66405232
#> 56  056  0.58155818  0.28219010
#> 57  057  0.13424691  1.30837423
#> 58  058 -1.29455904 -0.62962480
#> 59  059 -0.73977608  0.46959046
#> 60  060 -0.09516584  0.17431537
#> 61  061  0.09831719  0.66343260
#> 62  062 -1.71393157 -0.69836538
#> 63  063 -0.76902802 -1.05176726
#> 64  064  0.61445783  0.42611558
#> 65  065 -0.26574584  0.10461019
#> 66  066  0.52341509 -0.61624905
#> 67  067 -0.30231502 -1.56992438
#> 68  068  1.43703309  0.77068464
#> 69  069 -1.02472354 -2.23813125
#> 70  070  0.72061344  0.83243579
#> 71  071  2.34077309  1.73325542
#> 72  072  0.79658959  0.53480147
#> 73  073 -0.97082191  0.12676002
#> 74  074 -0.07171171  0.83673537
#> 75  075 -0.45004596 -0.88868064
#> 76  076  1.73091872 -0.70727048
#> 77  077  0.31310020 -1.00468305
#> 78  078  0.60638279  0.95727088
#> 79  079 -1.51172418 -1.89564856
#> 80  080  0.79011669  0.39913695
#> 81  081  0.02139684  0.73213238
#> 82  082  0.64030018  0.92919615
#> 83  083  0.60976597 -1.33426112
#> 84  084  0.13997614 -1.21709933
#> 85  085 -0.85832972  0.48984194
#> 86  086 -0.41878166  1.46292009
#> 87  087  0.17795315 -1.15699889
#> 88  088  1.65911772  0.25732700
#> 89  089 -0.10896503  0.62696295
#> 90  090 -0.48344121 -0.03061207
#> 91  091 -0.06362683 -0.50380718
#> 92  092  0.23982144 -0.53158000
#> 93  093 -0.59178470 -0.58642081
#> 94  094  0.82888041 -0.40879554
#> 95  095 -0.14651886  0.68609517
#> 96  096 -1.51571962 -0.35079824
#> 97  097  0.18326633  0.44856306
#> 98  098 -0.95300521 -0.29791104
#> 99  099  0.56360947  0.91702778
#> 100 100 -2.04879130 -0.72109224