Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4795189
#> Y_2 0.4795189 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -1.10721302 -0.5664050
#> 2 -0.19565151  0.7208232
#> 3 -0.20223642 -1.5163486
#> 4  0.14223548  0.4148604
#> 5  0.09628272 -0.2335686

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.161553475  0.473372111
#> 2   002  0.522552533  0.419318799
#> 3   003 -1.271749993 -1.066351270
#> 4   004  0.016449413  0.934007973
#> 5   005 -0.656829687  1.067502635
#> 6   006 -0.058933635 -1.940774555
#> 7   007 -0.208287398  1.998560156
#> 8   008 -1.271737915  0.181237785
#> 9   009  0.286414537 -0.709603232
#> 10  010  0.460173357 -1.276492930
#> 11  011  1.054606448 -0.442671354
#> 12  012 -0.751333105 -1.265910436
#> 13  013  1.378424295  3.534870160
#> 14  014  0.834360759 -0.278688301
#> 15  015 -0.930229274  1.170510114
#> 16  016  0.234284178 -0.296782233
#> 17  017  2.047769141  1.462514754
#> 18  018 -1.006089497 -0.525049621
#> 19  019 -0.002995539  0.085308724
#> 20  020 -0.533512620 -1.535412964
#> 21  021  1.262540533 -0.443293373
#> 22  022  0.681976782  0.324650564
#> 23  023 -0.017478511  1.605887005
#> 24  024 -1.092884211  0.320693736
#> 25  025 -1.000996822 -0.154349160
#> 26  026 -0.970443362  0.641119803
#> 27  027 -1.411454248 -0.651354469
#> 28  028  1.540339591  1.705366306
#> 29  029  0.569544229  1.371160200
#> 30  030 -0.148607545  0.119490446
#> 31  031  0.027915824  1.117987587
#> 32  032 -0.584989396 -1.760463321
#> 33  033  0.597655036 -1.734401744
#> 34  034  0.923369907  0.061871450
#> 35  035  2.432602221  2.699557410
#> 36  036 -0.481329804 -0.942082720
#> 37  037 -0.013709622 -0.554339953
#> 38  038 -0.105075742 -0.081063672
#> 39  039  0.341627385  1.318209537
#> 40  040 -1.329993192 -1.607129219
#> 41  041  1.386577310 -0.907140475
#> 42  042  0.419538159  0.211009304
#> 43  043 -0.093083351  0.556255518
#> 44  044 -0.034411311 -0.073147924
#> 45  045 -0.621488669 -0.960642627
#> 46  046 -0.019501435 -0.069153872
#> 47  047  0.573665019 -0.194350788
#> 48  048 -0.449983332 -0.354366703
#> 49  049 -3.389048165 -1.894392633
#> 50  050  0.972983361  1.169359142
#> 51  051  1.396730377  1.652466766
#> 52  052 -1.185149186 -0.134635890
#> 53  053  2.277495092  2.355770704
#> 54  054 -0.758519053  1.406660113
#> 55  055  0.361984273 -0.613850658
#> 56  056  0.304602331 -0.017391468
#> 57  057  1.277998496  0.036082065
#> 58  058  1.016634030 -1.231733146
#> 59  059 -1.634846439 -1.763588179
#> 60  060  0.044188163  0.819779452
#> 61  061 -0.200141728 -1.295251971
#> 62  062  1.253564609 -0.306177348
#> 63  063  0.423878062  1.112916954
#> 64  064  0.335033539 -0.740570064
#> 65  065  0.846000846 -0.791048466
#> 66  066  0.665261640  1.692091068
#> 67  067  0.809049163  0.156158617
#> 68  068 -0.633134189 -0.342637185
#> 69  069  0.189248907  0.227575394
#> 70  070  0.069845592  0.320508014
#> 71  071  2.336051867  1.156648598
#> 72  072  0.975329037 -0.561091979
#> 73  073  1.206786965  0.398311237
#> 74  074  0.791267553 -0.005481786
#> 75  075 -1.069670316 -0.032531396
#> 76  076  0.866388401  0.541873178
#> 77  077  0.554210741  0.806632290
#> 78  078  0.316241435  1.670945046
#> 79  079  0.075897687  0.472325599
#> 80  080  1.600576148 -0.037658385
#> 81  081 -0.691309101 -1.107159993
#> 82  082  0.130424482  0.426651136
#> 83  083  0.659684354  0.873308574
#> 84  084  0.807276572  0.187619260
#> 85  085  1.570594490  0.433475750
#> 86  086 -0.329674349 -1.168191865
#> 87  087  0.493075044  2.873015198
#> 88  088  1.636391254  1.186827013
#> 89  089 -0.096801983 -0.598862114
#> 90  090 -0.055536337  1.558078341
#> 91  091  0.700863416  0.744440803
#> 92  092 -0.580556314 -1.121961972
#> 93  093  0.409697524  0.240770791
#> 94  094 -0.354175753  0.739832568
#> 95  095 -1.595437762 -0.810567949
#> 96  096 -0.007281435  0.043081604
#> 97  097  0.225453596 -2.258642574
#> 98  098 -1.325415440 -0.450325567
#> 99  099  0.128959465  0.576270124
#> 100 100 -0.405457007  0.018158508

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.084687294  0.54374994
#> 2   002  0.768274534  0.77087348
#> 3   003 -1.696142523 -1.18566672
#> 4   004 -0.487363389  0.62701337
#> 5   005  0.237486611  0.12620879
#> 6   006  0.774639409  0.91830702
#> 7   007  0.667621919 -0.56164254
#> 8   008 -1.192179039 -0.84090470
#> 9   009  0.295079275 -1.72097719
#> 10  010 -1.213831183  1.39176923
#> 11  011  1.522734121 -1.09853971
#> 12  012 -1.100215332  0.04118424
#> 13  013 -0.874683311 -0.08757444
#> 14  014  0.345250941  0.48525805
#> 15  015  1.038236843  0.36919440
#> 16  016 -1.153159259 -1.90431784
#> 17  017 -0.742744498  0.19992016
#> 18  018 -1.127765224 -1.22533864
#> 19  019  0.206332645 -0.43581234
#> 20  020  0.053704471  0.46639700
#> 21  021 -0.958441304 -2.17216637
#> 22  022  0.057136563  0.76707357
#> 23  023  0.586241309  0.59151163
#> 24  024  0.755486733 -0.02751311
#> 25  025  0.791914773 -1.18405798
#> 26  026 -2.223696024 -1.47907574
#> 27  027 -0.070351099 -0.53123896
#> 28  028  0.241876581  1.21104602
#> 29  029 -0.455231787 -0.96492309
#> 30  030  0.110166279 -0.16090705
#> 31  031  2.085928214  1.72484407
#> 32  032 -0.648198224 -0.67314325
#> 33  033 -0.774499588 -0.45580167
#> 34  034 -0.005490631  0.18300350
#> 35  035 -0.871106490  1.77433880
#> 36  036  0.219781672  0.80239270
#> 37  037 -1.174561766 -1.53282237
#> 38  038  0.389789769  0.01034295
#> 39  039 -0.428368317  0.14084753
#> 40  040  0.357826847 -0.23128311
#> 41  041 -1.432826373 -1.93345644
#> 42  042  0.547386692  1.45591362
#> 43  043 -0.284079165 -0.60480286
#> 44  044  1.531083332  1.68091508
#> 45  045  0.659051776 -0.19335297
#> 46  046 -1.532338607  0.78708551
#> 47  047  0.962100937 -1.08541108
#> 48  048  0.033801829  0.15062014
#> 49  049  0.919411878  1.70024958
#> 50  050 -0.699542070 -0.65976696
#> 51  051  1.637761191  2.45308616
#> 52  052 -0.169547848  0.14616393
#> 53  053 -1.528857517 -1.41541448
#> 54  054 -1.562342029 -2.21531037
#> 55  055 -0.337525267  0.85964963
#> 56  056  0.716595262  1.02503438
#> 57  057 -0.083110271  0.19694741
#> 58  058  0.630915075  1.24090926
#> 59  059 -0.515343109 -1.00219449
#> 60  060  0.401154948 -0.89629634
#> 61  061  1.261346994  0.01005568
#> 62  062 -0.384587014 -0.08673570
#> 63  063 -2.040813112 -1.34034910
#> 64  064 -2.191445204 -1.38604480
#> 65  065  0.732365221  1.15431637
#> 66  066  0.623894369  0.05929403
#> 67  067  0.353178929 -0.22082074
#> 68  068  0.741010957  0.02198198
#> 69  069  0.749641687 -0.47984264
#> 70  070 -0.694872991  0.84953755
#> 71  071  1.255870156  1.89968793
#> 72  072 -1.931385189 -1.67571781
#> 73  073 -0.785784161 -0.64220579
#> 74  074  1.754651109  0.74893357
#> 75  075  1.190534694  0.37590755
#> 76  076  1.592067016  0.62434380
#> 77  077  0.780876611  1.07207549
#> 78  078 -1.553232499 -1.36832320
#> 79  079 -0.250746635 -1.65759820
#> 80  080  1.065516607  0.68665191
#> 81  081  0.142934534  1.24282413
#> 82  082  1.840731637  0.12344427
#> 83  083  0.296309290  0.95865251
#> 84  084  0.302748176  0.73888106
#> 85  085  1.380061244 -0.16137889
#> 86  086 -0.479441468 -0.40260705
#> 87  087  1.451238118  2.24489593
#> 88  088 -1.845450382 -1.25659534
#> 89  089  1.367281152 -0.03205181
#> 90  090  0.715087324  0.63890086
#> 91  091 -1.042884172 -0.67610034
#> 92  092 -0.797024990 -0.73762093
#> 93  093 -0.088705795 -0.20696129
#> 94  094 -0.563775580  0.59605093
#> 95  095  1.208625597  2.00434127
#> 96  096 -0.755609917  1.16422945
#> 97  097  1.342738655 -0.37089003
#> 98  098  0.612761112  0.45739656
#> 99  099  0.919891692  0.87819559
#> 100 100 -0.408065921  0.30478093