Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5301469
#> Y_2 0.5301469 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1       Y_2
#> 1  0.58242705 0.4987140
#> 2 -0.34197563 1.4832033
#> 3 -0.04634187 0.4745591
#> 4  0.98977451 0.9351023
#> 5  0.82840784 0.2257365

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.85173820 -0.71525604
#> 2   002 -0.09151114  0.42998279
#> 3   003  0.94418786 -0.79073637
#> 4   004  1.65936414 -0.34190888
#> 5   005  0.06852592  0.79433658
#> 6   006  0.49326024  0.86529040
#> 7   007  0.19500485  1.50148894
#> 8   008  2.40876831  0.62889403
#> 9   009 -0.15707966 -1.04193556
#> 10  010  0.60915869 -0.23818460
#> 11  011  0.23282829 -1.39799722
#> 12  012 -0.71987513  1.23906467
#> 13  013  0.09514041 -0.23452137
#> 14  014  0.05515183  0.42034899
#> 15  015  0.16803863  2.15146427
#> 16  016 -0.17904876  1.73736964
#> 17  017 -0.50787716 -0.30000296
#> 18  018 -0.50327407 -0.03108336
#> 19  019  1.29333595  0.27368026
#> 20  020 -0.27541955 -0.09963971
#> 21  021 -0.74590638  0.12932284
#> 22  022  0.01352351 -0.96250479
#> 23  023 -0.52564792  0.44515059
#> 24  024  0.82039665  0.21087365
#> 25  025 -0.28291622 -1.48083750
#> 26  026 -1.18059444  0.20995830
#> 27  027 -0.79988632 -0.33275389
#> 28  028 -3.23144918 -1.86221845
#> 29  029  0.50462939 -0.09609068
#> 30  030 -0.68683220 -0.04838591
#> 31  031 -1.04869578 -1.17827749
#> 32  032  0.42986186 -0.00873036
#> 33  033 -2.10741905 -0.63658244
#> 34  034 -1.40134328 -2.00074941
#> 35  035  1.81925101  2.46262276
#> 36  036  0.52769678 -0.82024735
#> 37  037 -0.15125217  0.33318689
#> 38  038 -0.09461749  0.69258327
#> 39  039 -0.38400916 -0.57731069
#> 40  040  0.31599586 -1.27330940
#> 41  041 -0.96264933  0.63218879
#> 42  042 -0.03407819 -0.28582985
#> 43  043  1.81420476 -0.47842553
#> 44  044  0.16331895 -0.13698189
#> 45  045  1.78763958 -1.10705740
#> 46  046  1.01664467 -0.13938633
#> 47  047 -2.35963231  0.63817014
#> 48  048  0.15691464  0.42287223
#> 49  049 -1.21126925 -0.49985823
#> 50  050  0.93572226 -0.43279395
#> 51  051 -0.23883130 -0.29096015
#> 52  052  0.36824926  0.58118040
#> 53  053 -0.67241644  0.15573465
#> 54  054  0.53755300  0.79783217
#> 55  055  0.65848633 -0.65931579
#> 56  056  0.63491597 -0.70860660
#> 57  057  0.58702646 -0.97289404
#> 58  058  0.52486322  0.30953388
#> 59  059 -1.40528216 -0.37626150
#> 60  060  0.51081130  0.65648603
#> 61  061 -0.99518742 -1.21297847
#> 62  062 -0.44885118 -0.19504555
#> 63  063  0.56059715 -1.08551373
#> 64  064  0.42304136  0.87235193
#> 65  065 -0.06491186 -0.96240113
#> 66  066 -0.28921721  1.19258554
#> 67  067  0.31246815  0.33273050
#> 68  068  0.96572167  0.24236131
#> 69  069  0.41812359  0.40989308
#> 70  070 -0.75974549  0.13430088
#> 71  071  0.74327279 -0.91065799
#> 72  072 -0.35277739 -1.20389441
#> 73  073  0.42384048 -0.81602113
#> 74  074 -0.33716294  0.43830105
#> 75  075 -0.13326204 -1.18029025
#> 76  076 -0.80838703  0.05742468
#> 77  077  2.48029592  1.97739978
#> 78  078 -1.12631547 -0.70398231
#> 79  079 -0.71134435  0.34853413
#> 80  080  0.99454442  2.26560486
#> 81  081 -0.32886336 -0.23140484
#> 82  082 -0.68532096  0.24580553
#> 83  083  1.31223004  1.88465342
#> 84  084  1.55690550 -0.24984169
#> 85  085 -1.35174276 -1.41391553
#> 86  086  1.25954039  1.36542129
#> 87  087 -0.28689470 -0.78963179
#> 88  088  1.47964102 -0.31267596
#> 89  089  0.22945742  1.69458871
#> 90  090  0.37237964 -0.25518661
#> 91  091 -0.20532651 -1.28143259
#> 92  092 -1.02602180 -1.81587437
#> 93  093 -1.74282165 -1.91349014
#> 94  094 -0.15473917  1.76852680
#> 95  095 -1.09273313  0.09077133
#> 96  096 -1.57108695  0.60090071
#> 97  097 -1.39948313 -1.07194008
#> 98  098 -1.15439553  0.75945506
#> 99  099  0.23440864  0.90963087
#> 100 100  0.84216273  0.84531224

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.66782985  0.984768448
#> 2   002  0.85058471 -0.350265362
#> 3   003 -1.10794556 -0.418228464
#> 4   004 -1.09355505 -0.329471099
#> 5   005 -0.57748138 -0.155553595
#> 6   006 -0.08047228 -0.585834765
#> 7   007 -0.53356574 -0.848261243
#> 8   008  0.17609151  0.046767663
#> 9   009  1.67810998 -0.144468403
#> 10  010  0.13257523  1.427118812
#> 11  011  1.04942130  0.657124111
#> 12  012  0.69948527 -0.464699077
#> 13  013 -1.65889332 -1.681631524
#> 14  014 -0.68626482 -0.050630601
#> 15  015 -0.61276286  0.387036513
#> 16  016 -0.80570037 -1.526234689
#> 17  017 -0.50703191 -0.417735276
#> 18  018 -2.63170082 -2.504444786
#> 19  019  0.20560470 -0.003238531
#> 20  020 -0.15455938  0.413125575
#> 21  021  1.61852526  1.607292399
#> 22  022  1.27553762  0.815059458
#> 23  023 -0.76320607  0.173072704
#> 24  024  0.80176630 -0.718659582
#> 25  025 -0.47429126 -1.568579984
#> 26  026 -0.23634391  0.565996336
#> 27  027 -1.11171971 -1.170476499
#> 28  028  0.98445895  0.513814383
#> 29  029 -0.21662047 -1.607356858
#> 30  030 -0.10590087 -0.314818330
#> 31  031 -1.24711640 -1.074157568
#> 32  032 -0.31479065 -1.072666443
#> 33  033  0.09691510 -0.443446274
#> 34  034  0.61086553  0.570848103
#> 35  035  0.49790338 -0.276206464
#> 36  036 -1.37989064 -2.253996020
#> 37  037 -2.15495845 -0.878516430
#> 38  038 -0.61167198 -1.399276904
#> 39  039 -0.16315001 -0.298459446
#> 40  040  0.56843691  1.177667627
#> 41  041 -0.01897529  0.286516849
#> 42  042 -0.83537812 -0.100850482
#> 43  043  1.50357639  1.177979733
#> 44  044  0.07430093  0.803127741
#> 45  045  1.01254909 -1.064013938
#> 46  046 -0.56961218 -1.080152729
#> 47  047  0.29445007  2.563837911
#> 48  048  0.62783045 -0.751734648
#> 49  049  0.27576180  0.477478601
#> 50  050  0.32213316  0.860696850
#> 51  051  1.14532044 -1.543443072
#> 52  052  0.64070139  0.999515172
#> 53  053  1.01230536 -0.813500196
#> 54  054 -0.75589117 -0.470887510
#> 55  055 -1.13007883 -1.358381085
#> 56  056  2.07453124  0.246164746
#> 57  057  1.61056017  1.423445729
#> 58  058  0.67616092 -0.518482852
#> 59  059 -0.21713054  0.276944080
#> 60  060 -1.91721830 -0.142411445
#> 61  061 -0.06668038  0.364460738
#> 62  062 -0.03826151  0.538795130
#> 63  063 -0.66352944 -0.307836281
#> 64  064 -1.96023314 -1.602805180
#> 65  065  0.50572298  1.202592934
#> 66  066 -1.19281395  0.502488340
#> 67  067  0.77899673 -0.633222703
#> 68  068  0.66101042 -0.610495904
#> 69  069  1.23428639  0.669523687
#> 70  070  0.42908216  0.484770195
#> 71  071  0.28651973 -0.105059446
#> 72  072  0.75036652  0.444018661
#> 73  073 -2.11718999 -0.623853214
#> 74  074 -1.38170483 -0.030387355
#> 75  075 -0.53561887 -1.255384570
#> 76  076 -0.20827208 -0.860537448
#> 77  077  0.16955493  1.240881383
#> 78  078 -0.11101740  0.393104306
#> 79  079  0.56580082  1.235032839
#> 80  080  0.06922764 -1.680730318
#> 81  081 -0.99816869  1.132128931
#> 82  082 -0.28528869  0.623269773
#> 83  083  0.81098548  0.073437888
#> 84  084 -0.65218318 -1.014429357
#> 85  085  0.37202798  1.393204249
#> 86  086  0.03562715  1.669486357
#> 87  087 -0.18924044  0.710809993
#> 88  088  0.32177007 -0.816672260
#> 89  089  2.47019514  2.103396094
#> 90  090  0.55550549  2.019318572
#> 91  091  0.78525851  0.307233721
#> 92  092  1.13239472  2.348246707
#> 93  093  0.36623195 -1.538067293
#> 94  094  1.60146916  0.582436932
#> 95  095  1.26361024  0.692649737
#> 96  096 -1.96315583  1.337055283
#> 97  097 -0.42321381  0.248308089
#> 98  098  0.07241267  0.385648891
#> 99  099 -0.10654737 -0.102027118
#> 100 100 -0.05802374  0.258082444