Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5028991
#> Y_2 0.5028991 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.8095941  1.0786792
#> 2  0.9736597 -0.6330146
#> 3 -0.7333386  0.3440604
#> 4 -0.8699178 -0.6945659
#> 5  0.4362008  0.6415505

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.78598814  0.253088498
#> 2   002 -0.70746820  1.198778243
#> 3   003 -0.60924987  1.071743665
#> 4   004  0.25770811 -0.141739191
#> 5   005  0.64258680  0.841918155
#> 6   006  1.99887663  0.215919039
#> 7   007  0.90057450  2.050964832
#> 8   008  0.41200181  1.956086637
#> 9   009 -0.98945519 -0.619189047
#> 10  010 -0.64108207  0.316852240
#> 11  011  2.50558653  1.330489458
#> 12  012 -1.37357555 -0.615122582
#> 13  013  0.46008741  0.022993629
#> 14  014 -0.64043846 -0.033744902
#> 15  015 -0.44432078 -1.233852459
#> 16  016 -0.06143587 -0.144479389
#> 17  017 -0.50295439  0.203944015
#> 18  018  0.26816219  0.437033666
#> 19  019  1.59351874  1.208440064
#> 20  020  0.09570358 -0.652391865
#> 21  021  0.77791053  0.846807162
#> 22  022 -1.07221931 -1.120516666
#> 23  023  2.13624085  0.476901545
#> 24  024 -1.69884027  0.004850673
#> 25  025  0.16426500  0.640099437
#> 26  026  0.08617387  1.160093156
#> 27  027  0.83495444  0.396278258
#> 28  028  0.32917674  0.842767899
#> 29  029  1.24892593 -0.958158013
#> 30  030  1.91939336 -0.812786189
#> 31  031  0.35959938 -0.127535612
#> 32  032 -0.46248945  1.046152204
#> 33  033  0.99422452  0.968897185
#> 34  034 -0.60992402 -0.909906266
#> 35  035  0.84883831 -0.178854934
#> 36  036 -0.81928184 -0.514913191
#> 37  037 -0.43790572 -0.286561150
#> 38  038  0.89499035  0.421892357
#> 39  039 -1.24084322 -1.754281403
#> 40  040 -0.08427483 -0.415451547
#> 41  041 -0.44272204  0.696950835
#> 42  042 -0.05665197  0.266104474
#> 43  043 -0.03802421 -1.599004695
#> 44  044 -0.93709434 -1.500216779
#> 45  045  0.81600936 -0.933108769
#> 46  046 -0.04476710  0.085163795
#> 47  047  1.10080601 -0.544204926
#> 48  048  1.51174593 -0.400759460
#> 49  049 -0.49103316  2.072251837
#> 50  050  0.65257440 -0.922963108
#> 51  051  0.63949085  1.514192250
#> 52  052 -2.52180985 -0.418688827
#> 53  053 -0.27830469  0.099253435
#> 54  054 -0.73705698 -1.329884633
#> 55  055  0.78375398  0.118597412
#> 56  056 -1.02385133 -0.320699574
#> 57  057 -0.51717771 -0.492192952
#> 58  058  0.75153696  0.449505575
#> 59  059  2.00807646  1.541495218
#> 60  060  0.41716164  2.126276868
#> 61  061 -0.41649403  1.212075502
#> 62  062 -1.66672112 -0.426035566
#> 63  063  0.66549943 -0.434077320
#> 64  064  1.31362222  1.200267999
#> 65  065 -0.08583515  0.346225769
#> 66  066 -0.10154131  0.106711796
#> 67  067 -0.23381775  0.463059975
#> 68  068 -1.68117398 -0.042050724
#> 69  069 -0.99211190 -3.636358024
#> 70  070  1.01283848  0.301583647
#> 71  071  0.43425701  0.078242692
#> 72  072  0.62777645 -2.348579447
#> 73  073 -0.07906643 -0.433822103
#> 74  074 -1.50894093 -0.932178086
#> 75  075 -0.16360789 -0.613224062
#> 76  076 -0.13646533  0.760864549
#> 77  077  0.32113686 -0.783148756
#> 78  078  0.23688469 -0.902912763
#> 79  079  1.42653987  1.605010077
#> 80  080  1.12349367  0.642241508
#> 81  081 -0.85131754 -2.730230775
#> 82  082 -0.50257191  0.026151552
#> 83  083  0.98689330 -0.425924621
#> 84  084  0.83216317  0.289686807
#> 85  085  1.43332352  0.706441119
#> 86  086  0.16666430  1.199068048
#> 87  087 -2.42162731 -1.995891110
#> 88  088 -0.09574103 -0.532072744
#> 89  089  0.25507833 -0.320666212
#> 90  090 -0.47099606 -0.321199393
#> 91  091  0.50956432  0.446060516
#> 92  092 -0.56287063  0.094310846
#> 93  093 -0.82385000 -0.427447577
#> 94  094  0.34987791 -0.896719192
#> 95  095  1.95351816  1.198316662
#> 96  096  0.13192739 -1.175549145
#> 97  097  0.35477409 -1.266520443
#> 98  098  1.24811074  0.152098236
#> 99  099  0.80386986  0.397474589
#> 100 100 -0.37710575  0.184223083

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.57490526 -1.00321692
#> 2   002  0.37649780  0.17147427
#> 3   003  1.77157391  0.83427161
#> 4   004 -0.11550443  0.62478999
#> 5   005 -0.98110539 -1.71293685
#> 6   006  0.32855250 -0.08448368
#> 7   007 -0.25300791 -0.49546789
#> 8   008  0.41301529  1.76185897
#> 9   009 -1.06942087  1.30073371
#> 10  010 -0.92727867 -1.16340386
#> 11  011 -1.26041462 -1.47851332
#> 12  012  0.97265102 -0.23549013
#> 13  013 -0.89750481  0.14285686
#> 14  014 -0.42881085  0.25441461
#> 15  015  1.19983131  0.93321837
#> 16  016 -0.48102719 -0.20083872
#> 17  017  0.02037028  0.16093515
#> 18  018  0.73410069  1.08948807
#> 19  019  2.05282253  1.61857171
#> 20  020 -0.18105966 -0.42692860
#> 21  021 -0.24815082 -1.00046317
#> 22  022  1.40241473 -1.51747544
#> 23  023  0.77788173  0.09248592
#> 24  024 -0.83219487 -0.74840015
#> 25  025 -0.59712962 -0.75326689
#> 26  026  0.21077476  1.63235240
#> 27  027 -1.04697023 -0.44806621
#> 28  028 -2.84476913 -1.61078031
#> 29  029  0.39361479  1.84552831
#> 30  030  0.51747526  1.23033194
#> 31  031 -2.39340620 -1.84918893
#> 32  032  0.21040699  1.31443356
#> 33  033 -1.33583186 -0.99706086
#> 34  034 -0.71211163 -1.18268477
#> 35  035  1.11393832 -1.92539987
#> 36  036  0.07332839 -0.82601671
#> 37  037 -1.03002192 -0.27936953
#> 38  038  0.27954345  1.10442149
#> 39  039 -0.10101270 -0.37200961
#> 40  040 -0.44344086  0.12335170
#> 41  041  0.90814467 -0.55586549
#> 42  042  0.52560049  0.20681375
#> 43  043  1.58335875  1.25878142
#> 44  044 -1.59131674  0.07238496
#> 45  045  0.66796044  0.77211898
#> 46  046 -1.06148846 -1.35341232
#> 47  047  0.28910507 -0.65196728
#> 48  048  1.02420725  1.25250053
#> 49  049 -0.05275066 -2.41076831
#> 50  050 -0.07357429  0.40736318
#> 51  051  0.19339716 -0.26638817
#> 52  052 -0.39292637  0.92797962
#> 53  053  0.27578615  0.84990345
#> 54  054 -0.91754725  0.35282569
#> 55  055  0.17082076 -0.11329781
#> 56  056  0.88418803 -0.41800435
#> 57  057 -1.13392066 -0.02419051
#> 58  058  0.62179411 -1.83703215
#> 59  059  0.98132330  1.11933201
#> 60  060 -0.19570571 -0.34360476
#> 61  061  0.29833694 -0.66842685
#> 62  062  1.44154876 -0.10241079
#> 63  063 -0.25195734  0.66396065
#> 64  064  0.15114039 -0.44235508
#> 65  065 -0.13673147  0.04239387
#> 66  066 -0.78806607 -0.86888396
#> 67  067  0.16888588  0.41467945
#> 68  068 -0.58845452  0.07476763
#> 69  069 -1.79370161 -1.12531648
#> 70  070 -0.50454553 -1.49798019
#> 71  071 -0.97643438 -0.18866804
#> 72  072  1.34129801  1.42933854
#> 73  073  0.14747328 -1.27766368
#> 74  074  1.13100403  0.02937050
#> 75  075  0.30688676  0.20942369
#> 76  076  0.10085096  0.23497495
#> 77  077  0.36624241  0.93351391
#> 78  078 -0.92163954 -0.97921545
#> 79  079  0.30311693  1.38252699
#> 80  080 -0.71631235 -1.91376227
#> 81  081  0.40816987 -0.64518618
#> 82  082  1.64672775 -0.01948042
#> 83  083 -1.14335670 -0.72086481
#> 84  084 -0.28715050  0.98511596
#> 85  085 -2.34292793 -0.74165655
#> 86  086 -0.95998859 -1.23459350
#> 87  087  0.52503307 -0.30065419
#> 88  088  0.37167012  1.42484749
#> 89  089  0.85017243  1.08250954
#> 90  090  0.67854773  2.52058544
#> 91  091  0.27152139  0.80028305
#> 92  092  0.41001609  1.43008060
#> 93  093 -0.67506336  0.91237601
#> 94  094 -0.29822369  0.25235992
#> 95  095 -1.10223660 -1.40307710
#> 96  096  0.35228620  0.25496839
#> 97  097  1.59912447 -1.32032866
#> 98  098 -0.91094444 -1.55019954
#> 99  099  1.15557658  0.34871137
#> 100 100  0.60264926 -0.93724540