Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5051229
#> Y_2 0.5051229 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1          Y_2
#> 1  0.6457073 -0.007379053
#> 2  1.0479773  1.902109776
#> 3  0.6927718 -0.855051326
#> 4  2.1315544  1.147320897
#> 5 -0.4062409 -0.547066423

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.08621896  1.15035079
#> 2   002  1.47536490  0.71830963
#> 3   003  0.64464849  0.29117428
#> 4   004 -0.05979529  0.23822468
#> 5   005 -0.39388297 -0.35780410
#> 6   006 -1.52414547 -0.98572609
#> 7   007  0.30555116  0.15450608
#> 8   008  0.12605147  0.40790477
#> 9   009 -0.33136214 -0.13423644
#> 10  010 -1.21448396  0.34470605
#> 11  011 -0.54315959 -0.07180711
#> 12  012 -0.11501032 -0.84982577
#> 13  013  0.02003455 -0.83279591
#> 14  014 -0.10036389 -0.89063877
#> 15  015  1.06080274 -0.45869077
#> 16  016 -0.18736715 -0.68751420
#> 17  017 -0.74189609 -0.65587306
#> 18  018 -0.55170458  0.17221555
#> 19  019  2.21397186 -0.13304898
#> 20  020  0.93844155 -1.04238546
#> 21  021  2.23791361  1.28125214
#> 22  022  0.99582792  2.58177805
#> 23  023 -0.09280589  1.45601494
#> 24  024  1.42383722  0.43953708
#> 25  025 -0.04076012  1.01266771
#> 26  026 -0.58919440  0.41966792
#> 27  027 -1.70562220 -2.18636561
#> 28  028 -0.95058691 -0.77321694
#> 29  029 -0.05066132 -1.16208579
#> 30  030  0.48471544 -1.15163028
#> 31  031  0.53608641  0.51307957
#> 32  032 -0.09620571  0.48393746
#> 33  033  0.22988458  0.29189757
#> 34  034  0.82215264 -0.02660365
#> 35  035 -0.16461897 -1.22164839
#> 36  036 -0.12123321  0.56436546
#> 37  037  0.85713466  0.84793955
#> 38  038 -0.09014240  1.69160288
#> 39  039 -0.93838535  0.34087105
#> 40  040 -0.13117675  0.95371563
#> 41  041 -1.42701941 -1.25284756
#> 42  042 -1.06730815  0.91203071
#> 43  043 -0.92830158 -0.21941620
#> 44  044  0.30345131 -0.88652520
#> 45  045 -0.32094008 -1.09618064
#> 46  046 -0.77520634  0.06648666
#> 47  047 -0.43280031  0.69033984
#> 48  048 -2.41557501 -1.29000010
#> 49  049  0.12854291  1.31478893
#> 50  050  0.15312176 -0.44433182
#> 51  051 -0.57006960  0.12606696
#> 52  052 -2.05185430 -2.58109314
#> 53  053  0.57509492 -0.93250024
#> 54  054  0.56987329 -0.02831355
#> 55  055 -0.24332547 -0.14039326
#> 56  056  1.12363588  1.29263507
#> 57  057  0.37912369  2.94224180
#> 58  058  0.16686580  0.92147491
#> 59  059 -0.66171017 -0.68349595
#> 60  060 -1.36816534 -1.98610746
#> 61  061 -1.16914541 -2.48161060
#> 62  062  1.32095374  0.60232754
#> 63  063 -0.93200974 -0.51250456
#> 64  064  1.64710580  1.11496316
#> 65  065  0.11141344 -0.44283893
#> 66  066 -0.54588844  0.55024846
#> 67  067 -0.25878767  0.09845346
#> 68  068 -0.46281960 -0.65715687
#> 69  069  0.83163154  0.73936478
#> 70  070  1.14251810  1.57158972
#> 71  071  0.66570020  2.37622922
#> 72  072 -0.31103729 -0.39074348
#> 73  073  0.69602284 -0.11578815
#> 74  074  0.22772026 -0.47506696
#> 75  075  1.08438756 -0.04059523
#> 76  076  1.46551315  2.39050637
#> 77  077  0.21431363  0.18606162
#> 78  078  0.43075856  0.66379122
#> 79  079 -0.33586457  0.79202032
#> 80  080 -1.21072009 -0.53531201
#> 81  081 -0.86745752 -0.06738635
#> 82  082  1.00572435  1.79566065
#> 83  083 -0.39373350 -0.11152800
#> 84  084  0.71883341 -0.47534642
#> 85  085 -1.26258099  0.91946975
#> 86  086 -1.77887769 -0.06667979
#> 87  087 -1.46322928 -1.75598833
#> 88  088 -1.11198194 -0.51746085
#> 89  089  0.93540444  1.28264627
#> 90  090  0.71051173  1.86316798
#> 91  091  1.43860966 -0.32150047
#> 92  092 -0.23962866 -0.15751624
#> 93  093 -0.16107350 -0.03144739
#> 94  094  1.09963453  1.91223439
#> 95  095 -1.21005883 -1.34886743
#> 96  096 -1.72195994 -1.78012735
#> 97  097  0.44498536  0.16816145
#> 98  098  1.64254179  1.57721081
#> 99  099  0.51691194  0.42032712
#> 100 100 -1.72474069 -0.15304588

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.531253006  0.66937103
#> 2   002  0.659240367  0.76779772
#> 3   003 -0.021997860  0.15594341
#> 4   004  0.272461203 -1.54413569
#> 5   005  1.497164221  0.73266934
#> 6   006 -0.759413590 -0.26552481
#> 7   007 -1.087163617  0.19323298
#> 8   008  1.552909546  1.99842844
#> 9   009  0.094922842  0.55452808
#> 10  010 -1.558875917 -1.71770057
#> 11  011  0.995460168  0.26588882
#> 12  012 -1.122091347 -0.90305736
#> 13  013 -1.398890193 -1.42780382
#> 14  014 -0.204593447 -0.07421132
#> 15  015  0.732833067 -0.74100293
#> 16  016  0.093117534 -0.18901994
#> 17  017 -0.089444937 -0.97898810
#> 18  018  0.151689390  1.04257929
#> 19  019  1.270777818 -0.16762444
#> 20  020  0.788421185  0.37291984
#> 21  021 -0.047337777  0.72491529
#> 22  022 -1.021462694 -0.58427318
#> 23  023  0.005840611 -1.19768691
#> 24  024  0.352160989 -0.85254249
#> 25  025  0.610385937  0.58746838
#> 26  026  1.037452527  0.46999344
#> 27  027  1.172908287  1.16278670
#> 28  028  0.526419324 -0.60332097
#> 29  029 -0.799581877 -0.40737099
#> 30  030  1.140712230  0.81952604
#> 31  031  0.864898699  0.21439867
#> 32  032 -0.663174818 -0.03781832
#> 33  033  0.450333999  0.47860974
#> 34  034  0.785424618 -0.90791785
#> 35  035 -0.341358314  1.25642607
#> 36  036 -1.684531037 -1.10719994
#> 37  037 -0.639191456 -0.08457011
#> 38  038 -0.591573926 -0.73617366
#> 39  039  0.827450237  0.24334448
#> 40  040  0.081705131  0.96239550
#> 41  041 -1.161120318  0.26365587
#> 42  042  0.008243372  0.90663837
#> 43  043  1.103793038 -0.24406794
#> 44  044  0.960676271  0.11432905
#> 45  045  1.307547095 -0.04369145
#> 46  046 -0.562444535 -0.74534676
#> 47  047 -0.182653743 -0.19457352
#> 48  048 -0.102989528 -0.13151818
#> 49  049 -0.370573673 -0.64183090
#> 50  050 -0.871173932 -0.58558402
#> 51  051  1.111763482 -0.14420303
#> 52  052 -1.286213163 -1.62699390
#> 53  053 -0.605351675 -1.60158409
#> 54  054  1.734323256  0.33198966
#> 55  055  2.060885138  1.22941593
#> 56  056 -0.544102160 -1.26205141
#> 57  057  0.289252136 -0.08142098
#> 58  058 -0.724390321  1.19459244
#> 59  059 -0.927898250 -0.40253068
#> 60  060 -0.944262097  0.23710443
#> 61  061 -0.459839751 -1.20866207
#> 62  062  0.163000754  1.09267876
#> 63  063 -1.115205002 -0.58417063
#> 64  064 -1.261836792 -1.64521898
#> 65  065  0.702915520  0.60981570
#> 66  066  0.821772618  1.54274094
#> 67  067 -0.817911026 -0.84059472
#> 68  068  0.670929665 -1.00467369
#> 69  069  1.222505909 -0.70937446
#> 70  070  0.995777030 -0.77479813
#> 71  071 -0.745990920 -0.90387012
#> 72  072 -0.473615028  1.00091756
#> 73  073  1.801360263  1.59742539
#> 74  074  0.414393675  1.31420141
#> 75  075 -0.176398745  1.11445778
#> 76  076 -0.221666749  1.38250253
#> 77  077  0.134436143  0.62535270
#> 78  078  0.573528329  1.74009094
#> 79  079  1.128864484  0.91926906
#> 80  080  1.821263190  1.48788919
#> 81  081 -0.362117819 -0.70182590
#> 82  082  1.110938454  1.77654503
#> 83  083  0.929218638  0.94343528
#> 84  084  1.210356931  0.07749729
#> 85  085  0.781782181  0.99432080
#> 86  086  0.836000781  1.15514485
#> 87  087 -0.984022021 -0.45251492
#> 88  088 -0.308592315 -0.47708362
#> 89  089 -1.004691256  0.05197181
#> 90  090  1.210088448  0.62945601
#> 91  091 -0.428766826 -0.17945487
#> 92  092 -2.677560742 -1.08536863
#> 93  093 -1.189498964 -0.41650590
#> 94  094 -0.136106581 -0.26136295
#> 95  095  0.312914285  1.34706284
#> 96  096  0.716034578  1.09406772
#> 97  097  0.584807027 -2.09863652
#> 98  098 -1.362433283 -0.68656037
#> 99  099 -0.522966498 -0.64702826
#> 100 100  1.777295803 -0.17615681