Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.547231
#> Y_2 0.547231 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.1765536  0.4277405
#> 2  1.6413260  1.6732460
#> 3 -0.4967526 -0.2251673
#> 4  0.6101149 -0.1770678
#> 5 -0.8559280  0.4412486

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.072777368  0.09091021
#> 2   002  1.652206967  0.56689242
#> 3   003  0.176540581 -1.06456816
#> 4   004 -1.483242010  0.15145365
#> 5   005 -1.187119814 -1.56634662
#> 6   006 -0.560583098 -0.59503624
#> 7   007 -0.975394338 -0.78970927
#> 8   008  0.441158460 -0.59342197
#> 9   009 -0.061364900  0.61786570
#> 10  010  0.611098141  1.00812223
#> 11  011 -0.119488197 -0.69279649
#> 12  012  2.140304658 -0.14267772
#> 13  013 -1.177156195 -0.68671929
#> 14  014 -1.435163913 -2.45514858
#> 15  015  0.053877990  0.95331264
#> 16  016  0.377169095  0.60997843
#> 17  017 -0.548728891 -0.50320903
#> 18  018 -1.599006044  0.63695656
#> 19  019  0.438738663 -0.31779337
#> 20  020 -0.048106022  0.63668517
#> 21  021  0.985282477  0.87096944
#> 22  022  0.994552203 -0.44214064
#> 23  023 -0.008292446 -0.11205519
#> 24  024 -1.064413742 -0.39256693
#> 25  025  0.240305436  0.84777232
#> 26  026 -0.535280409 -0.04704659
#> 27  027  0.627002338 -0.73517560
#> 28  028 -1.207389511  0.42718112
#> 29  029  0.081004638  0.55772087
#> 30  030  0.514476385  0.62902177
#> 31  031 -0.163719506  0.03860429
#> 32  032 -1.436876374 -0.50428689
#> 33  033  0.615113052  0.49234976
#> 34  034  0.200976594 -0.90207118
#> 35  035  0.094018477  0.05199293
#> 36  036  1.288309904  0.80442077
#> 37  037  0.817291267  0.05613205
#> 38  038  0.776912617  1.24216265
#> 39  039  0.217643009 -0.27235595
#> 40  040  1.354124591 -0.03608829
#> 41  041  0.177938497 -0.08671998
#> 42  042  3.149744039  1.70857999
#> 43  043 -0.808727149 -0.40611205
#> 44  044 -0.888183894  0.16794650
#> 45  045 -0.286588733 -0.58738810
#> 46  046  1.750845681 -0.05575193
#> 47  047 -0.121506316 -0.53024138
#> 48  048 -0.283159255 -0.58277320
#> 49  049 -1.325471377 -1.84474357
#> 50  050 -0.044718749  0.22771922
#> 51  051 -1.383245403  0.25871212
#> 52  052  0.580946525  2.20933660
#> 53  053 -0.379314164  0.31444620
#> 54  054  1.020492377  0.34121813
#> 55  055  0.207178526  0.57207615
#> 56  056 -0.840591330 -0.15638736
#> 57  057  1.806112744  1.91923758
#> 58  058  1.494007802  1.36149873
#> 59  059 -0.253783953  2.27742765
#> 60  060 -0.497393073 -0.97785895
#> 61  061  0.956608315  1.29276455
#> 62  062 -0.874815519 -0.57112222
#> 63  063  0.085408739  0.53127085
#> 64  064  2.034604159  1.28903143
#> 65  065 -0.278211535 -0.51245200
#> 66  066  0.944240086  1.30237789
#> 67  067 -0.117813574 -1.30363936
#> 68  068  1.263161622 -0.42920629
#> 69  069 -0.454109765 -1.20232041
#> 70  070  0.494353921  0.52907609
#> 71  071  0.837556455  0.96459093
#> 72  072 -0.985568000  0.78023999
#> 73  073 -0.515669214  0.28651029
#> 74  074 -1.084355575 -0.47019793
#> 75  075  0.390349292  0.21392073
#> 76  076  2.543854378  1.68804230
#> 77  077  0.618841437  0.87736397
#> 78  078  0.304099553 -0.42875573
#> 79  079  1.226229737 -0.15353765
#> 80  080  0.111820595 -1.06954380
#> 81  081 -0.207761269  1.06025778
#> 82  082 -0.134351912 -1.11049099
#> 83  083 -0.026669679 -0.47151816
#> 84  084 -0.200176280 -0.47998474
#> 85  085  0.504485052  0.76991244
#> 86  086 -0.896613074 -0.66561733
#> 87  087 -0.090386566 -1.59199741
#> 88  088 -0.035754109  0.26315728
#> 89  089 -1.364779320 -0.10933707
#> 90  090  0.911367475 -0.44532102
#> 91  091 -1.207680230  0.04928108
#> 92  092  0.913760561  1.80649029
#> 93  093  0.248915483 -2.82138354
#> 94  094  0.252445408 -0.36777489
#> 95  095  1.506734242  2.71543851
#> 96  096  0.007746918 -0.38997900
#> 97  097  1.243598271  1.03607234
#> 98  098  0.478646842  0.78791365
#> 99  099  1.810663730  0.93424260
#> 100 100  0.472965886 -0.29011703

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.323723911  2.064308333
#> 2   002  0.237002905  0.719430014
#> 3   003  0.685560436  1.212228066
#> 4   004 -0.119312429 -0.101469640
#> 5   005  0.526999013 -0.899668120
#> 6   006  0.938020096  1.005108973
#> 7   007  0.109580779 -0.254920518
#> 8   008  0.155438634  0.820623477
#> 9   009  1.548692631  1.017420272
#> 10  010  1.292464546 -0.152123250
#> 11  011 -0.370602135  0.095772853
#> 12  012 -1.734647153  0.154741747
#> 13  013 -0.141854611  2.111037513
#> 14  014 -1.430261550 -1.486477256
#> 15  015 -0.786850307 -1.761549445
#> 16  016 -0.355747790  0.527094354
#> 17  017 -0.607546164 -0.111599935
#> 18  018  1.281418435 -0.940333071
#> 19  019 -1.002625639 -1.637893527
#> 20  020 -1.115925930 -1.284611907
#> 21  021 -1.289435059 -0.489293886
#> 22  022  1.509852635  0.502680571
#> 23  023 -0.006522963  0.755889055
#> 24  024  0.361310390  0.626147771
#> 25  025  1.882039853  2.494072361
#> 26  026 -0.313418324  1.071233391
#> 27  027 -0.805234375 -1.755963871
#> 28  028  0.331507219 -0.179199719
#> 29  029  1.271522848  1.579074407
#> 30  030  1.829621661  1.978385389
#> 31  031  0.445996379  0.511586628
#> 32  032 -1.462939255 -0.529916186
#> 33  033 -1.981861109 -1.603326507
#> 34  034  0.606075976 -0.176092513
#> 35  035 -0.971714883  0.004967837
#> 36  036  0.604077347  0.930774934
#> 37  037  0.899546060 -0.571283351
#> 38  038 -1.348828512 -0.929461997
#> 39  039 -0.355974799 -1.110631970
#> 40  040  0.230927868 -0.321291280
#> 41  041  0.678279138  1.133132918
#> 42  042 -1.796913084 -1.552067799
#> 43  043  0.585922027  0.072752103
#> 44  044  0.279335184  0.336665712
#> 45  045  0.788419379  1.276632474
#> 46  046  1.883359983  0.274355786
#> 47  047 -0.506433411  0.616304097
#> 48  048  1.012474005  1.128521566
#> 49  049 -1.039036408  0.067704524
#> 50  050 -0.126294224  0.630501746
#> 51  051  0.316628266  1.049556752
#> 52  052  1.037614635  0.195945385
#> 53  053 -0.321663647 -0.556624131
#> 54  054  1.037986046  1.095735742
#> 55  055 -2.270991948 -0.113812706
#> 56  056  0.689366461  0.064429381
#> 57  057  0.588333760 -0.112625488
#> 58  058  0.655328977  1.853695302
#> 59  059  0.007237456 -0.145612353
#> 60  060 -1.973521765 -0.841849853
#> 61  061  2.132865938  0.011036888
#> 62  062 -1.618913254 -0.506082129
#> 63  063 -0.200506050  0.164078147
#> 64  064 -2.170747700 -0.987191506
#> 65  065  1.463734260  1.867144288
#> 66  066  1.895330938  0.232570395
#> 67  067  1.302757933 -0.482633415
#> 68  068 -0.341379386  0.596519918
#> 69  069 -0.249510647  0.222021302
#> 70  070 -0.329148954  0.431080756
#> 71  071 -0.930269137 -0.832853886
#> 72  072 -1.238335037  0.451209010
#> 73  073 -0.336171134 -0.524404364
#> 74  074  0.351326168 -1.330570613
#> 75  075 -0.255933896 -1.476449238
#> 76  076  0.021314601 -1.477981482
#> 77  077  0.393338795 -0.159883040
#> 78  078  0.353666935 -0.326135114
#> 79  079 -1.486725408 -0.208919967
#> 80  080 -0.091143411 -1.189514854
#> 81  081 -0.026518638  0.320169388
#> 82  082  0.412100227  0.793155284
#> 83  083 -1.249472895 -0.236822591
#> 84  084 -0.522417826 -0.148404490
#> 85  085  0.751947949  0.442298729
#> 86  086 -1.004891221  0.062547858
#> 87  087  1.640274840  2.130533438
#> 88  088 -0.324855999  0.483336442
#> 89  089  0.581876747  1.398582095
#> 90  090  0.917818648 -0.018822340
#> 91  091 -0.587963608  0.066243723
#> 92  092  1.406884007  0.455740513
#> 93  093 -1.309219576 -0.957276304
#> 94  094 -0.567144907 -0.514701108
#> 95  095  1.128091862  0.567628166
#> 96  096 -1.494684582 -0.995476161
#> 97  097  0.137793566 -0.274577417
#> 98  098  0.669779269  0.247134579
#> 99  099 -0.666205243  0.704623372
#> 100 100 -0.039197109 -1.415366626