Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5143503
#> Y_2 0.5143503 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.46470125 0.60972774
#> 2 -0.07320471 0.23548875
#> 3  1.02827819 0.01214625
#> 4  0.80798874 0.16780764
#> 5  0.34121734 0.61824672

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.67269266  0.24041832
#> 2   002 -0.09513592 -0.91535441
#> 3   003 -0.45772009 -1.50022228
#> 4   004  0.60489458  1.23647646
#> 5   005  0.26566209 -0.13155322
#> 6   006  1.11041845  2.23739819
#> 7   007  0.33848544 -0.60992473
#> 8   008  0.69703529  0.68497146
#> 9   009  0.31496224  0.42309397
#> 10  010  0.56159738  0.94788670
#> 11  011 -0.82736036  0.53844630
#> 12  012  0.85365095 -0.48764677
#> 13  013 -0.43322984  0.98078921
#> 14  014  1.46663729  0.84943441
#> 15  015 -0.22354426 -1.04351706
#> 16  016 -1.24315731 -0.45172183
#> 17  017  0.19952702  0.93710530
#> 18  018  1.34025094  0.59894911
#> 19  019  0.01497109  1.32794286
#> 20  020 -1.65614239 -0.76033608
#> 21  021 -1.01337528 -1.42756776
#> 22  022  0.69999933 -0.32150284
#> 23  023 -1.12979863 -1.46154477
#> 24  024 -0.19859709  0.05017315
#> 25  025 -0.59763880  1.23036583
#> 26  026  0.80708283  1.29874767
#> 27  027 -1.00806463 -1.77948447
#> 28  028  0.86472618  0.08800909
#> 29  029  1.02941852 -0.08838326
#> 30  030 -0.88465092  0.37070461
#> 31  031  1.32228934  0.18710695
#> 32  032  0.47268994  0.03366427
#> 33  033 -2.62788633 -1.07519020
#> 34  034  0.05554436 -1.35206194
#> 35  035 -1.76284299 -1.77343917
#> 36  036 -0.41257342 -0.53285807
#> 37  037  0.16667958 -0.15662845
#> 38  038  0.36697308  0.44879825
#> 39  039  0.34358368 -0.68677124
#> 40  040  1.24946856 -0.43176329
#> 41  041 -0.48573705 -0.76827104
#> 42  042  0.11847640 -0.73197948
#> 43  043  0.77293274  1.42840993
#> 44  044 -0.06200699 -0.77827317
#> 45  045  0.08071634  0.39443463
#> 46  046 -0.34261300  0.56589132
#> 47  047 -0.36314813  1.98220481
#> 48  048  1.88601301  1.30181674
#> 49  049  0.86784729  0.62771651
#> 50  050 -1.45291138 -1.26910781
#> 51  051 -0.75291412  0.27763286
#> 52  052 -0.05690546 -1.26299729
#> 53  053  0.94301637  0.15017237
#> 54  054 -0.64164192  0.13899544
#> 55  055  0.64947676  1.58392660
#> 56  056 -0.29762891  0.38453774
#> 57  057 -0.29170344 -0.91075412
#> 58  058 -0.87208494 -0.99493780
#> 59  059 -0.27825351  0.21222756
#> 60  060  0.60065981  0.54351424
#> 61  061  0.82348633 -0.67227342
#> 62  062 -1.02908430 -0.14044325
#> 63  063  1.03883619 -0.39383229
#> 64  064  0.45333666 -1.25185767
#> 65  065 -0.79284279  0.42124153
#> 66  066  0.30993294 -0.09067223
#> 67  067  0.98816191 -0.09535775
#> 68  068 -0.20509642 -1.46105797
#> 69  069  0.38056978 -1.16647949
#> 70  070  0.88539469  1.11062821
#> 71  071 -0.38449881  0.34561222
#> 72  072  0.35974448  0.59796781
#> 73  073 -0.20150345 -0.76712389
#> 74  074 -0.25936485  0.75878785
#> 75  075 -0.60441492 -1.11189713
#> 76  076  0.44067154 -0.06455986
#> 77  077  0.44486280  0.41099298
#> 78  078 -0.73060065 -1.01956232
#> 79  079 -0.23501762  0.06347256
#> 80  080 -2.43327365 -1.07791401
#> 81  081  0.64013653  0.16276205
#> 82  082 -0.97436649 -0.79761628
#> 83  083 -0.45296828  0.40580748
#> 84  084  0.05736869 -1.32881498
#> 85  085 -0.79521267  0.18969249
#> 86  086  0.84105536  1.76451063
#> 87  087  0.43282295 -1.09873724
#> 88  088 -0.33858651 -2.07533938
#> 89  089 -0.66957203  0.39987016
#> 90  090  0.93758342 -0.59904926
#> 91  091 -0.45422261  0.17447002
#> 92  092 -0.15017075  0.55216143
#> 93  093 -0.02254764 -1.01095891
#> 94  094 -1.70467901 -1.22749199
#> 95  095  1.30082177  1.29222519
#> 96  096 -1.47878648  0.94400429
#> 97  097  0.24303461 -0.13426597
#> 98  098 -0.86868877  0.09949034
#> 99  099 -0.72406886 -0.27422043
#> 100 100 -0.90809819 -1.91005396

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.23783880 -1.35454563
#> 2   002  1.39054897  0.23294129
#> 3   003 -0.73471640 -1.18290330
#> 4   004  1.07275298  0.19226600
#> 5   005  0.71997464  1.48094500
#> 6   006  0.34517789  0.67109927
#> 7   007  1.16054240  1.96098618
#> 8   008 -0.28169710  1.26226836
#> 9   009 -0.53511546 -0.57958774
#> 10  010 -1.53987514 -1.34680740
#> 11  011  0.87159639  1.20002067
#> 12  012 -1.31998970 -2.10983804
#> 13  013 -0.15348238 -1.41084970
#> 14  014  1.25718952 -0.66901635
#> 15  015  1.65360697  1.23895923
#> 16  016  0.21710600  0.23679966
#> 17  017  0.44850846  0.59847838
#> 18  018  0.32600767 -0.42884844
#> 19  019 -0.54041012 -1.17528509
#> 20  020 -0.68760985 -0.19658085
#> 21  021  2.66601066  1.40193596
#> 22  022 -1.11917217 -1.53706306
#> 23  023  0.91756495 -1.72071949
#> 24  024 -0.57248553  0.57601312
#> 25  025 -0.13643119 -0.67291298
#> 26  026  0.07422048  1.42701402
#> 27  027 -0.99333174  0.31926695
#> 28  028 -1.95232621 -0.28082588
#> 29  029 -2.70846597  0.62553181
#> 30  030 -1.07399624 -2.58706384
#> 31  031 -1.01389086 -0.76384266
#> 32  032  0.14982028  1.39361065
#> 33  033 -0.33747355 -1.17130560
#> 34  034  0.14962237  1.72399598
#> 35  035 -1.84332078 -0.42218800
#> 36  036 -1.12631573 -2.38312466
#> 37  037 -0.92289801  0.37051567
#> 38  038 -0.92737967 -0.62034994
#> 39  039 -1.10964323 -1.36217882
#> 40  040 -0.07785701  0.78308922
#> 41  041  0.81942872 -1.17819541
#> 42  042 -1.22962595  0.17815699
#> 43  043  0.28101602 -0.61269566
#> 44  044  1.05118319 -1.52322479
#> 45  045 -0.15927770  0.97777535
#> 46  046 -0.26020446  0.53627342
#> 47  047  1.31789766  0.13355539
#> 48  048  0.43424912 -0.25830830
#> 49  049  1.29194455 -0.42788103
#> 50  050  0.98669568  0.08065047
#> 51  051 -0.63212991  0.20693992
#> 52  052 -0.89356150 -0.21536824
#> 53  053  0.62625705  0.94611670
#> 54  054  0.79809507  0.46178491
#> 55  055  0.84883486 -0.49179567
#> 56  056  0.95655810  0.22436143
#> 57  057  1.56441974  1.45237659
#> 58  058 -1.14491027 -0.09631848
#> 59  059 -0.96604339 -1.63556396
#> 60  060  0.64609404 -0.37456758
#> 61  061  1.47896666  0.72909152
#> 62  062 -0.05128470 -0.16326941
#> 63  063 -0.71596971  0.97839434
#> 64  064  0.06817008  1.27499162
#> 65  065  0.18877985  1.22257148
#> 66  066  0.66429608  0.70693702
#> 67  067  0.61944530  1.04185841
#> 68  068 -0.82692731 -2.06573583
#> 69  069 -0.83993550  1.25051800
#> 70  070  0.01152928  1.13177531
#> 71  071 -1.26293270 -0.59648743
#> 72  072 -1.38620675  1.47710596
#> 73  073 -0.71688732  0.10222741
#> 74  074 -1.29400633 -0.81077683
#> 75  075  0.80172020  1.83049844
#> 76  076  0.17521746 -2.36761063
#> 77  077  0.28281055  1.41407013
#> 78  078  1.05365938  0.11042774
#> 79  079  0.61066106  0.39622853
#> 80  080  0.78334977 -0.39655159
#> 81  081  0.54279257  1.41178122
#> 82  082 -0.10209773  1.62956120
#> 83  083 -2.23738677 -2.58263518
#> 84  084 -0.58098189 -1.19240489
#> 85  085 -0.29141730 -0.29652052
#> 86  086 -0.09664774 -0.26414025
#> 87  087  0.67097922  0.91626580
#> 88  088 -0.54724597 -0.43873870
#> 89  089 -1.06108491 -1.00292920
#> 90  090 -0.68168888 -0.72860960
#> 91  091  0.11957219 -0.43933762
#> 92  092 -0.58480667 -1.99105832
#> 93  093 -1.22790868 -1.27517078
#> 94  094 -0.79677131  1.54351902
#> 95  095 -0.71108822  0.09788499
#> 96  096 -0.11161452 -0.79918432
#> 97  097  0.67865705 -0.29498642
#> 98  098 -0.63420881 -0.37029618
#> 99  099 -2.39523900 -0.64487651
#> 100 100 -0.09827619 -0.65151454