Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5164496
#> Y_2 0.5164496 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.22060858 -0.7196910
#> 2  0.76257067 -0.9390973
#> 3  0.10035665  2.0708702
#> 4  0.11639499  0.2168039
#> 5  0.00302525  0.5878071

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.54014681 -2.05716673
#> 2   002  0.66279780  0.02818755
#> 3   003  0.08896198  1.17828032
#> 4   004  1.37053552 -1.47341268
#> 5   005 -1.21225121 -1.00367249
#> 6   006 -0.49959560 -1.04990636
#> 7   007 -1.49300572 -0.67885381
#> 8   008  0.58347587 -1.55102648
#> 9   009  1.42283982  1.39410962
#> 10  010  1.20781316  0.75862317
#> 11  011  0.11732623 -0.32716511
#> 12  012  1.45955115  1.26186412
#> 13  013  0.05922652  1.82168171
#> 14  014  0.81077990 -0.26117729
#> 15  015  0.28226888  0.90356576
#> 16  016 -0.81590762 -1.50237521
#> 17  017 -1.03039235 -1.10570550
#> 18  018 -0.03227561  0.87178863
#> 19  019  0.16693592  0.86635744
#> 20  020 -1.54775519 -0.84106524
#> 21  021  1.36510089  0.92874769
#> 22  022 -0.62355865 -2.14516140
#> 23  023 -0.07080937 -1.29577406
#> 24  024  1.43703801 -0.40304177
#> 25  025 -0.66876481 -1.62898759
#> 26  026 -0.02772735  0.47886589
#> 27  027 -1.05753471 -0.67043719
#> 28  028  1.08552192  1.63632375
#> 29  029  0.99781494  0.02127834
#> 30  030  0.64049473 -0.23832856
#> 31  031 -1.22424631 -0.90463902
#> 32  032  2.17929208  2.02192379
#> 33  033 -0.61802757  0.43399268
#> 34  034  1.42105975 -0.14592138
#> 35  035  0.03401514 -0.09824973
#> 36  036 -1.70839771 -0.86911407
#> 37  037  0.61635013  1.10779467
#> 38  038 -0.77862143 -0.07044573
#> 39  039 -1.70024787  0.53228308
#> 40  040 -1.03907373 -0.39979286
#> 41  041 -0.74298087 -1.33844983
#> 42  042 -1.62103067 -0.96878167
#> 43  043 -0.47916736  0.92267343
#> 44  044  1.86365557  0.59004470
#> 45  045 -0.45323005  0.09427415
#> 46  046 -0.38251523  0.41173979
#> 47  047  0.31547748 -0.74644381
#> 48  048  0.07284677 -0.88071739
#> 49  049 -1.10292042 -0.02559678
#> 50  050  2.57848703  1.47122486
#> 51  051  0.17301495 -1.11580576
#> 52  052 -0.14410747  0.28382420
#> 53  053  0.04262029 -0.48484755
#> 54  054  0.50894265 -1.29197072
#> 55  055  1.60060012  1.23267529
#> 56  056 -0.54622565 -1.01361053
#> 57  057 -0.50624189  0.56913594
#> 58  058  1.45839806  2.35141544
#> 59  059 -0.15518072 -1.61883567
#> 60  060 -0.86790144  0.30990873
#> 61  061  0.15616689  0.77952262
#> 62  062  0.75119754  1.51059497
#> 63  063 -0.61616021 -1.45639099
#> 64  064 -0.50901032  0.31538002
#> 65  065  0.45691343 -0.45061884
#> 66  066  1.30814140  0.10380835
#> 67  067 -0.69181074 -0.99512761
#> 68  068  0.31618409  0.41681877
#> 69  069  1.59810261  0.23411168
#> 70  070 -0.32283299 -2.30364747
#> 71  071  1.14579371  0.07259133
#> 72  072  0.23931105  0.05617438
#> 73  073 -1.48354119 -1.37952280
#> 74  074  1.13124716  0.01536096
#> 75  075  0.86590260  1.00117864
#> 76  076  0.09297527 -0.13704308
#> 77  077 -0.56698544 -0.83082888
#> 78  078  1.46073622 -0.74620796
#> 79  079  0.43667627 -0.47216605
#> 80  080 -0.36933415  1.48975932
#> 81  081 -0.07850240  0.44880833
#> 82  082  0.06213030  0.11948336
#> 83  083  0.26448899  0.79038896
#> 84  084  1.13868466  1.82731579
#> 85  085  0.65461890  0.74711845
#> 86  086 -0.55927161 -1.33134818
#> 87  087 -0.21631240 -0.25847735
#> 88  088  0.88069178  1.27509664
#> 89  089 -0.76328463 -1.57801520
#> 90  090  1.29857331  0.66189508
#> 91  091 -0.35450625  2.08216580
#> 92  092 -1.09983000  0.47844261
#> 93  093  0.84165789  0.02398804
#> 94  094 -0.88607097 -2.35377543
#> 95  095  1.67367686 -0.55369412
#> 96  096  0.52774219  2.00543973
#> 97  097  0.44286874 -0.47401322
#> 98  098 -0.15723836 -1.75807417
#> 99  099  2.07240188  1.19622065
#> 100 100 -0.81429567 -0.18174211

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001 -3.769570e-01 -0.72328299
#> 2   002  4.676010e-01 -0.16708003
#> 3   003 -3.086973e-01 -0.13081319
#> 4   004 -1.517104e+00 -0.81046342
#> 5   005 -9.395965e-01 -0.37761303
#> 6   006 -5.265609e-02 -0.21333530
#> 7   007  8.021360e-01  0.52852541
#> 8   008  4.217667e-01  0.10303878
#> 9   009  1.492046e+00  0.04604333
#> 10  010 -1.160162e+00 -1.00346394
#> 11  011 -9.221744e-01  0.16918076
#> 12  012  8.086366e-01  0.21315128
#> 13  013 -1.117939e+00 -0.30243600
#> 14  014  1.917528e-01  1.45749356
#> 15  015 -8.120562e-01 -1.36468559
#> 16  016 -6.616049e-01  0.67366525
#> 17  017  9.142710e-01 -0.02860706
#> 18  018  2.030602e+00  0.54948899
#> 19  019 -1.334384e+00  0.54361358
#> 20  020  7.624787e-01  1.27965078
#> 21  021  6.840821e-01  1.09548257
#> 22  022 -2.356516e+00 -0.50334880
#> 23  023 -1.898390e-01 -0.01989239
#> 24  024 -5.260467e-01  0.52151758
#> 25  025 -1.176176e+00  0.45657632
#> 26  026 -1.897883e-01  0.42476872
#> 27  027 -1.290680e+00 -0.02880661
#> 28  028  6.633731e-02  1.15453688
#> 29  029 -9.201649e-01  0.15395903
#> 30  030 -3.287934e-05 -0.14955541
#> 31  031 -1.134998e+00 -1.70506551
#> 32  032 -3.358332e-01  0.36328955
#> 33  033  5.081492e-01  1.85637683
#> 34  034  8.202517e-01 -0.21976732
#> 35  035 -1.247679e+00 -1.91049663
#> 36  036  1.083935e+00  0.13191548
#> 37  037  3.446331e-01  0.80187431
#> 38  038 -1.989784e-01 -1.59562365
#> 39  039  7.126634e-01 -1.49267284
#> 40  040 -1.134225e-01  0.61743876
#> 41  041  1.277444e+00  0.07265228
#> 42  042 -5.877321e-01 -0.16627996
#> 43  043 -3.818807e-01 -0.65392883
#> 44  044 -6.118199e-01 -0.01883420
#> 45  045 -1.362450e-02 -0.04202566
#> 46  046 -8.584354e-01 -0.78398471
#> 47  047 -2.336281e+00 -1.12865234
#> 48  048 -1.839618e-01 -2.43095735
#> 49  049 -1.613985e+00  0.31688244
#> 50  050 -9.807619e-01 -1.57291243
#> 51  051 -1.230760e+00 -1.16830715
#> 52  052 -4.646147e-01 -0.03266450
#> 53  053 -1.488824e+00 -0.91229531
#> 54  054  2.530705e+00  0.02016530
#> 55  055  3.115240e-01  0.63734513
#> 56  056 -6.545957e-01 -0.00877128
#> 57  057  1.648856e+00  0.70130389
#> 58  058 -1.081613e+00 -1.77709087
#> 59  059 -1.381225e+00 -1.41998548
#> 60  060 -1.434851e-02 -0.92915696
#> 61  061  9.161630e-01  1.24147539
#> 62  062  1.003728e+00  1.29095632
#> 63  063  4.019331e-02  1.29667540
#> 64  064  9.121820e-01 -1.04096547
#> 65  065  1.892613e-01  1.69246166
#> 66  066 -5.921403e-01  0.56427108
#> 67  067 -2.533060e-01  0.47226088
#> 68  068 -1.672584e+00 -0.89599724
#> 69  069 -3.154774e-02  0.63153488
#> 70  070 -1.727236e+00 -0.88761155
#> 71  071 -4.687429e-01  0.87564436
#> 72  072 -6.586252e-01 -1.62110432
#> 73  073 -1.276332e-02  0.50279913
#> 74  074  2.460976e+00 -0.16394529
#> 75  075 -9.287385e-01 -0.02480187
#> 76  076  3.382635e-02  0.36817602
#> 77  077  1.643185e+00  0.76408377
#> 78  078 -1.142960e+00  1.47361836
#> 79  079  8.601896e-01  0.46001692
#> 80  080 -1.133130e-01  0.14895621
#> 81  081  7.904459e-01  1.09020175
#> 82  082 -7.709294e-01 -1.02874022
#> 83  083 -1.237162e+00  0.39322280
#> 84  084  1.023815e+00  1.08460521
#> 85  085  1.520427e-01  0.77213352
#> 86  086 -7.626325e-01  0.60541813
#> 87  087 -1.323229e-01 -1.70651857
#> 88  088 -2.654484e-01  0.07069032
#> 89  089 -5.598668e-01 -0.16580382
#> 90  090 -4.156282e-02  0.36066716
#> 91  091  1.686828e+00  0.25065283
#> 92  092  1.930655e-01  0.47630746
#> 93  093 -2.642129e-01 -1.18088236
#> 94  094 -1.098525e-01  1.92386699
#> 95  095 -5.280870e-01  1.15026741
#> 96  096  2.921860e-01 -0.30667205
#> 97  097  4.402098e-02 -0.76781191
#> 98  098  1.546932e-01  0.20426266
#> 99  099  1.063687e+00  1.72763304
#> 100 100  5.787785e-01  0.47093613