Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4552433
#> Y_2 0.4552433 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.9235535 -1.65344303
#> 2  1.4045175 -0.62322855
#> 3 -0.8972552 -0.86427565
#> 4 -1.8529207 -0.75299131
#> 5 -0.9913900 -0.07007257

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.22671481  1.340148802
#> 2   002  0.28685395  0.062875074
#> 3   003  0.04663859  0.361072288
#> 4   004  2.33552609  2.638201626
#> 5   005 -3.20957165 -0.719949011
#> 6   006 -0.15721505  0.579953096
#> 7   007 -0.43853298  0.361483343
#> 8   008 -1.32009893  0.410742822
#> 9   009 -1.15009700 -0.552322150
#> 10  010  0.80266815  1.283511261
#> 11  011  0.65786830 -0.611449517
#> 12  012 -1.97069021 -2.257821373
#> 13  013  0.02049878  0.536242087
#> 14  014  0.01393573 -0.240861927
#> 15  015  0.98369874 -1.523566247
#> 16  016  0.40314635  1.494072077
#> 17  017 -0.45870916 -0.558807393
#> 18  018 -0.38658716 -0.428139858
#> 19  019 -0.73300238 -0.534285954
#> 20  020  0.90187242  0.967690539
#> 21  021 -0.90725893 -2.130498590
#> 22  022 -0.20540931 -0.546346591
#> 23  023  1.26204521  1.662695527
#> 24  024  1.66796376  0.512361865
#> 25  025  1.22150388 -0.680486483
#> 26  026 -0.33154587  0.140995899
#> 27  027 -0.53103938  1.154401921
#> 28  028 -1.08010020 -0.619692651
#> 29  029  0.72906942  0.450287890
#> 30  030  0.75280599 -1.173846399
#> 31  031  0.40928152  0.903728151
#> 32  032  1.35459798 -0.963025055
#> 33  033 -0.33084977 -0.638497888
#> 34  034 -0.19894803  0.602410287
#> 35  035  1.03334219  1.153333030
#> 36  036 -0.14402042  0.596894099
#> 37  037  0.46830413 -0.819197031
#> 38  038 -0.06999750 -0.071947299
#> 39  039 -1.00183217 -1.936587517
#> 40  040 -0.67125875 -0.884932246
#> 41  041  1.34598177  2.294695393
#> 42  042  1.16161656 -0.367265572
#> 43  043 -0.34743288 -0.374846921
#> 44  044  1.27194844  0.657575641
#> 45  045  0.23419138 -0.393479969
#> 46  046 -0.79864615 -0.634408541
#> 47  047 -1.40229683  0.291433557
#> 48  048  0.54133038 -0.643542105
#> 49  049  0.34073517  0.932786372
#> 50  050  1.58298954  1.076952082
#> 51  051  0.31446631  0.626422131
#> 52  052 -1.39218903  1.597078614
#> 53  053 -0.19591096  0.109977852
#> 54  054 -0.47276619  1.074008652
#> 55  055 -0.56396890  1.811432606
#> 56  056  2.28846961  1.402763792
#> 57  057  0.79246322  0.637670551
#> 58  058  0.40515776  1.784874356
#> 59  059 -0.41366649 -1.761070158
#> 60  060  1.12347877  0.620173892
#> 61  061  0.66128165 -0.450322152
#> 62  062 -0.55883026 -1.713291617
#> 63  063 -1.01487044  0.432284100
#> 64  064  1.29149317  0.696849484
#> 65  065 -1.90026454 -0.679682942
#> 66  066  0.04117037  0.741242224
#> 67  067 -0.21206002 -0.176235149
#> 68  068  0.95093525  2.033836900
#> 69  069 -0.76580209 -0.332278700
#> 70  070  1.50654904  0.897757345
#> 71  071 -0.92472643 -0.210041463
#> 72  072  1.73563871 -0.478527942
#> 73  073  0.05996605 -0.555957423
#> 74  074  0.03346575  1.400394221
#> 75  075  1.46224993  0.910100315
#> 76  076  0.30384668 -0.700721290
#> 77  077 -0.44089662  0.489900637
#> 78  078  1.11116200  1.299382815
#> 79  079  0.56171828  0.226247834
#> 80  080 -1.14904354 -1.243556148
#> 81  081 -0.06790284 -0.248248396
#> 82  082  1.02372500  0.515558679
#> 83  083  0.08885818  1.109854146
#> 84  084 -0.82160228  0.015010880
#> 85  085 -0.18136722 -0.330203494
#> 86  086  1.23204460  0.815789738
#> 87  087  0.06126468 -0.005303801
#> 88  088  2.07756769 -0.366854563
#> 89  089 -1.04495253  0.120952969
#> 90  090 -0.58177246  0.334820662
#> 91  091  1.43889230  0.019840355
#> 92  092  0.47784144  1.246686850
#> 93  093  0.34829390 -0.506648695
#> 94  094  0.26143188 -0.694228891
#> 95  095  0.13769807 -1.715133724
#> 96  096  0.07326852  0.713829467
#> 97  097 -0.73795874  0.228999987
#> 98  098  0.43823826  0.140854941
#> 99  099  1.59430344  2.025119679
#> 100 100  1.07365495  1.481753191

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.059748998 -0.45044934
#> 2   002  2.398173556  2.43150679
#> 3   003 -0.844183548  1.28417134
#> 4   004  0.550176032 -1.02737362
#> 5   005  0.498856888  1.06268032
#> 6   006 -0.362257719  0.23348257
#> 7   007 -1.103351484 -1.25935438
#> 8   008 -0.117093234 -0.66376281
#> 9   009  0.134328911  0.73868111
#> 10  010 -0.650400183  0.65301873
#> 11  011  0.080495888 -0.87838809
#> 12  012 -0.559569201 -0.20853294
#> 13  013  2.084005173  0.46506668
#> 14  014 -0.331352699  0.44683056
#> 15  015  0.175944311  1.46496478
#> 16  016  1.688173557  0.45669747
#> 17  017  0.556591505 -0.96188130
#> 18  018  0.536920852  1.20253239
#> 19  019 -0.306062759  0.50359834
#> 20  020  0.463962436  0.61344905
#> 21  021  1.061141758  1.05919679
#> 22  022  1.718184993  0.47476332
#> 23  023  0.060286854  0.59462554
#> 24  024  1.609812706  0.53712882
#> 25  025  0.301598660  0.44835986
#> 26  026  1.143651920  1.81163825
#> 27  027 -0.550236378  2.04950657
#> 28  028  0.562402666  0.81378700
#> 29  029 -0.530389453  1.93536720
#> 30  030  0.469095965 -0.13397293
#> 31  031 -1.812451732 -2.37264973
#> 32  032  0.273197288 -0.87099459
#> 33  033  1.453706062  1.77134093
#> 34  034  0.446076070  1.61242490
#> 35  035  0.583073085 -0.64639758
#> 36  036  0.457390019  0.30513516
#> 37  037 -2.727381033 -1.47515886
#> 38  038  0.647003280  0.94952704
#> 39  039  0.655674696  0.85783014
#> 40  040  0.764794365 -0.17059978
#> 41  041  1.669823029  0.40743799
#> 42  042  0.762509338 -1.28064886
#> 43  043 -0.431172694 -0.62096291
#> 44  044 -0.031099954  1.05772134
#> 45  045 -1.349103024 -1.09227377
#> 46  046  0.793522423  0.52728140
#> 47  047  0.299492335  1.14759020
#> 48  048  0.826894291  0.01567982
#> 49  049  0.093640564 -0.34927480
#> 50  050  2.296959243  0.95726902
#> 51  051 -0.539780209  0.12208721
#> 52  052 -0.357794585  0.15163753
#> 53  053 -0.600016184 -1.42849235
#> 54  054 -1.009946512 -0.27307808
#> 55  055  0.863165027 -0.77342117
#> 56  056  0.854725132  1.78566927
#> 57  057 -0.743662053  0.11778681
#> 58  058  1.704563818  0.35002513
#> 59  059 -1.831960696 -1.36388956
#> 60  060 -0.050525519  1.36561663
#> 61  061  1.435853061  1.75822933
#> 62  062  0.055170639 -1.48983022
#> 63  063  1.410130849  0.64202919
#> 64  064 -1.008322412  0.35380138
#> 65  065  1.270742594  0.01921609
#> 66  066  0.286548831  0.79367432
#> 67  067  1.198433388 -0.53573906
#> 68  068 -0.792399371 -0.75094534
#> 69  069  1.233996506  0.35719337
#> 70  070 -0.352386213 -2.17293542
#> 71  071 -0.270505279 -1.02146277
#> 72  072 -0.762571911 -0.21669183
#> 73  073 -0.752993560 -0.14035453
#> 74  074 -0.061855955  0.53634328
#> 75  075 -2.061470191 -0.73229699
#> 76  076  0.913768115  0.71168030
#> 77  077  1.989197159  1.99577408
#> 78  078 -0.154620020 -0.16036569
#> 79  079  0.847375730  0.14737785
#> 80  080 -1.419307367  0.31253010
#> 81  081 -0.819906100 -0.17857122
#> 82  082  0.227635171  0.95912927
#> 83  083  0.108056874 -0.65494685
#> 84  084 -0.770422195  0.10531794
#> 85  085  1.106954479  1.87903042
#> 86  086 -0.863303268 -1.22828526
#> 87  087 -0.224210156  0.36524472
#> 88  088  0.636065221  2.13928757
#> 89  089  0.470713069  0.60371767
#> 90  090 -0.204315924 -0.78087513
#> 91  091 -0.716120422  0.19493538
#> 92  092  1.127016153  1.86700605
#> 93  093  0.993507712  0.10029843
#> 94  094 -1.428938037 -0.21682726
#> 95  095 -1.041030777 -1.69743386
#> 96  096  0.654543337  1.74147072
#> 97  097 -0.004149131 -0.69465845
#> 98  098 -0.651946416 -1.68454499
#> 99  099  0.943695930 -1.30651877
#> 100 100 -0.934871489 -0.53776589