Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4666402
#> Y_2 0.4666402 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1 -0.50920866  0.51233722
#> 2 -0.06812919 -0.62236587
#> 3 -0.36065091 -0.57406728
#> 4 -0.53685445 -1.22063621
#> 5  0.69144738 -0.05594304

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.633691142  0.40656711
#> 2   002 -0.525975514  0.23887401
#> 3   003 -1.011208089  0.13573771
#> 4   004 -0.020549586  0.47596222
#> 5   005 -1.255336451  0.63063731
#> 6   006  0.270553095 -1.34867487
#> 7   007 -0.439438826 -1.23153342
#> 8   008  0.400239016 -2.19391907
#> 9   009 -1.002430683 -0.54806701
#> 10  010 -1.123765511 -0.07873440
#> 11  011  0.824768296  0.57499401
#> 12  012 -0.095475480  0.92028245
#> 13  013  1.414474652 -0.17954997
#> 14  014  0.565453875  1.92280044
#> 15  015  0.405814372 -0.22422758
#> 16  016 -0.008169423  2.04857509
#> 17  017  2.366381405  0.66321737
#> 18  018  1.488805678  0.38334087
#> 19  019 -0.211707309  1.15042919
#> 20  020 -0.874782106 -0.04972861
#> 21  021 -0.515458516 -0.59223079
#> 22  022  0.179776799 -1.35768513
#> 23  023  0.588629799  0.04521369
#> 24  024 -1.186428059 -0.28229001
#> 25  025 -0.064450743  1.30812619
#> 26  026 -0.608833765  0.80726122
#> 27  027  0.524691757 -1.55877089
#> 28  028 -0.384705599  1.13945681
#> 29  029  0.247321052 -1.15989976
#> 30  030  0.440873190  1.06496357
#> 31  031 -0.088593668  0.26039684
#> 32  032  0.215775916  0.83945475
#> 33  033 -1.014447569 -1.58548154
#> 34  034  1.011100482  0.21226923
#> 35  035 -0.161595577 -0.90040395
#> 36  036 -0.107149088  0.78034452
#> 37  037  0.715458449 -1.20384927
#> 38  038  1.344505907 -0.50235466
#> 39  039 -0.835162161  0.61888378
#> 40  040 -0.210277475 -1.56872880
#> 41  041  0.149399828  1.52618072
#> 42  042 -0.500906848 -0.22965684
#> 43  043 -0.881156353 -1.63692179
#> 44  044  0.308955598  1.32331419
#> 45  045  0.139931876  0.69400192
#> 46  046  0.227346207 -0.78804059
#> 47  047 -0.590477512 -0.39115389
#> 48  048 -0.712081728  0.81448025
#> 49  049  0.740041573  0.69090248
#> 50  050  1.078847795  1.37108049
#> 51  051  0.487696308  1.24671914
#> 52  052  0.330863286 -0.69252414
#> 53  053  1.245700486  0.67827158
#> 54  054 -0.974577755  0.09732944
#> 55  055 -0.062347429 -0.11417240
#> 56  056 -1.042117116 -2.12651442
#> 57  057 -0.409215104 -1.23729532
#> 58  058  0.835792356  0.54768695
#> 59  059 -0.321019147  0.85013703
#> 60  060  0.241971851 -2.06808862
#> 61  061 -0.196862977 -0.79004093
#> 62  062 -0.132863741  0.51602011
#> 63  063  0.081105127  0.42314649
#> 64  064  0.429905459 -1.65687450
#> 65  065 -0.202566555 -0.97705573
#> 66  066  1.732432814  2.13081645
#> 67  067  1.747758664  2.24516311
#> 68  068 -0.158790343  0.86180842
#> 69  069 -3.117103209 -1.08706661
#> 70  070  0.319044514  1.48879182
#> 71  071 -0.394755897 -0.54551391
#> 72  072  0.217176301  0.33235633
#> 73  073  0.381127554  1.03213765
#> 74  074  1.578029565  2.12101085
#> 75  075 -0.648673019 -0.43194362
#> 76  076 -0.114927063  1.55092168
#> 77  077  0.307307176 -1.16947598
#> 78  078 -1.110372604 -0.76002777
#> 79  079 -0.243951287 -1.54929236
#> 80  080  0.728499870 -0.21469153
#> 81  081 -0.588796068  0.24421460
#> 82  082 -0.753543320  0.33065499
#> 83  083  0.322236620  1.04324487
#> 84  084 -1.671080922 -1.17712376
#> 85  085  0.257524811  1.68482205
#> 86  086 -0.390146057  0.49483516
#> 87  087  0.452862888 -1.15249129
#> 88  088  0.817264859  0.71263054
#> 89  089 -3.201960006 -2.36318065
#> 90  090  1.047070556  0.26451282
#> 91  091  0.991805789 -0.01741493
#> 92  092  0.259299925  0.77825295
#> 93  093 -2.305425138  0.49555908
#> 94  094 -0.727097115 -1.08994046
#> 95  095  0.814627012  0.44751216
#> 96  096  1.437964805 -0.38657491
#> 97  097  1.232073528  0.44473617
#> 98  098 -0.155638221  0.98748451
#> 99  099  0.396357552 -0.24390908
#> 100 100  0.142389125 -0.58250431

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.06683180  0.597827142
#> 2   002 -0.26706876 -0.729421304
#> 3   003  1.80055009  1.630079296
#> 4   004 -2.54718695 -0.164464476
#> 5   005  0.87634447  0.855761426
#> 6   006  1.12588722  1.132481588
#> 7   007 -1.12369531 -0.248153752
#> 8   008 -0.01064480  0.291873475
#> 9   009  0.84379606 -0.480348092
#> 10  010  0.39461140  0.749836612
#> 11  011 -0.90052308 -0.603575726
#> 12  012 -0.93355278 -2.171005730
#> 13  013  0.19261040  0.709133889
#> 14  014  0.38770375  0.950174685
#> 15  015  0.66393957  1.535067095
#> 16  016 -1.24120487 -0.998860737
#> 17  017  1.20096824  1.183614626
#> 18  018  0.76593958  0.472445813
#> 19  019 -1.04426515 -0.921770811
#> 20  020  0.60704731  0.545519719
#> 21  021  0.50951907 -0.589958422
#> 22  022 -0.65088605  0.068512618
#> 23  023  0.65445255  0.872679521
#> 24  024  0.05262499  2.317728691
#> 25  025  0.78408092 -0.731143306
#> 26  026  1.37544753  0.605965949
#> 27  027 -0.70245681 -0.631325285
#> 28  028  0.92019609  0.630649134
#> 29  029 -1.05148248  0.189127520
#> 30  030  1.22135363  0.992624683
#> 31  031  0.26019526 -0.282345004
#> 32  032 -0.95683229  0.822163702
#> 33  033 -0.07822310  0.221678225
#> 34  034  1.26225767 -1.599246561
#> 35  035  1.81043648  0.891197323
#> 36  036 -0.19223565  0.489283087
#> 37  037  0.50209965 -0.685055984
#> 38  038  0.56362492 -0.661895080
#> 39  039  1.37034808  1.146135604
#> 40  040 -0.72229970 -0.832596613
#> 41  041 -0.08607419  1.172020388
#> 42  042  1.22436734  1.882730850
#> 43  043 -0.35486401  1.008796447
#> 44  044  1.50027518  0.689299758
#> 45  045  0.63175228 -0.347754995
#> 46  046  0.01277525  0.899553245
#> 47  047 -1.83497837 -1.590515277
#> 48  048  1.34199288  1.534164827
#> 49  049 -0.56243083  0.549081711
#> 50  050 -0.08140761 -0.605230636
#> 51  051 -2.61226977 -2.596327882
#> 52  052  0.14026235 -0.503726000
#> 53  053  0.72352510  0.678707568
#> 54  054  2.12141870  0.667466278
#> 55  055 -2.19005245 -0.551636791
#> 56  056  0.43146487  0.201988016
#> 57  057 -0.73255666  0.458924789
#> 58  058  0.72768461 -0.415256744
#> 59  059 -0.55188340  0.731707363
#> 60  060  0.23564256  0.715636926
#> 61  061  0.52353883  0.591355555
#> 62  062  0.14510668 -0.053877297
#> 63  063  0.66024827 -0.230062353
#> 64  064 -1.66930062 -0.802629023
#> 65  065  0.73382649 -1.064568864
#> 66  066  0.79603842  1.411086326
#> 67  067  1.08391618  0.744136201
#> 68  068  0.67745767 -0.350752397
#> 69  069 -0.36046097  0.204253607
#> 70  070 -1.55261966 -0.372794341
#> 71  071  0.17854758 -0.051734730
#> 72  072 -0.88793929 -0.973004792
#> 73  073 -0.77415343  0.125373569
#> 74  074 -1.39462992 -0.549764556
#> 75  075  0.73784961  1.345316244
#> 76  076 -0.13931430 -1.103755068
#> 77  077 -2.36397042 -1.658731061
#> 78  078 -1.46115890 -2.136143886
#> 79  079  1.52720528  1.229971903
#> 80  080  0.39884693 -1.561726943
#> 81  081  1.02625138  1.083345660
#> 82  082 -0.12570381  1.784561092
#> 83  083 -0.08145266 -0.002747784
#> 84  084  1.11920757 -0.459588984
#> 85  085  0.95396722  0.699014027
#> 86  086  0.52187840  0.155779617
#> 87  087 -0.08551410 -0.684688107
#> 88  088  0.30243872 -0.433079471
#> 89  089  1.59078789  0.048506495
#> 90  090  0.14396240 -0.192142961
#> 91  091  1.21669651  0.962520361
#> 92  092 -1.80998876 -0.653268268
#> 93  093  0.48845840 -1.107981426
#> 94  094 -1.70816545 -1.859525048
#> 95  095  1.38928622  1.798202721
#> 96  096  1.04173903 -0.979782744
#> 97  097 -0.21206572 -0.541295038
#> 98  098  0.55931503  0.408996889
#> 99  099 -0.32672749 -0.024461969
#> 100 100 -0.31390405 -0.971415050