Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.521596
#> Y_2 0.521596 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.09742169  1.4153737
#> 2 -1.55380043  0.3337384
#> 3 -0.91233550 -0.7096485
#> 4 -1.23627139 -0.3411587
#> 5 -1.87470207 -0.3628656

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.75197668  0.846452311
#> 2   002 -0.31895771 -0.625541579
#> 3   003  0.83327394  0.370745197
#> 4   004  1.01976229  0.883854823
#> 5   005  0.15435952 -0.349227769
#> 6   006  0.89135622  1.420669275
#> 7   007  0.78202810  0.245238971
#> 8   008  0.94407077  0.643672904
#> 9   009  1.50367162  0.190619794
#> 10  010 -1.48866091  1.355899218
#> 11  011 -0.63149792 -0.901445247
#> 12  012 -0.97920470  0.553247906
#> 13  013 -0.73117757  1.711330105
#> 14  014  0.30795831  1.025579526
#> 15  015  0.66607710  0.122927861
#> 16  016  0.83437923  0.027161374
#> 17  017 -0.83561465 -0.424312610
#> 18  018  0.27755832  0.360911528
#> 19  019  0.48850844 -0.884018353
#> 20  020 -0.53573546 -1.079854059
#> 21  021 -0.65159635  0.262886795
#> 22  022 -1.30565594 -0.515437578
#> 23  023 -0.87432583  0.017274058
#> 24  024 -1.19495943 -1.547601761
#> 25  025 -1.54186920 -0.871110702
#> 26  026  1.53115302 -0.665843599
#> 27  027 -0.38492276 -0.352479485
#> 28  028  0.32053625 -0.513131062
#> 29  029  0.04213967 -0.112678488
#> 30  030  1.16341078  0.854127448
#> 31  031 -0.04562293 -1.947729058
#> 32  032 -0.07070758 -1.036670735
#> 33  033 -1.08671698  0.134723158
#> 34  034  0.57844511 -0.164189746
#> 35  035  0.56128171  1.225661427
#> 36  036 -0.73277017  0.337931627
#> 37  037  1.43112373  1.244431823
#> 38  038  0.73942252  0.211191555
#> 39  039  0.81731985  0.038184716
#> 40  040  0.55933541  1.044456595
#> 41  041 -0.52369327  0.145105793
#> 42  042  2.04543706  0.946550374
#> 43  043  0.18901533  1.006052801
#> 44  044 -0.68761907 -0.889558102
#> 45  045 -0.49106963  0.931244505
#> 46  046  0.36267651 -0.230328034
#> 47  047 -1.05777935 -0.838583897
#> 48  048 -0.92188000  0.007732437
#> 49  049 -0.92620067 -1.288631630
#> 50  050  0.06915521 -0.053137429
#> 51  051 -1.19798539 -1.518640715
#> 52  052 -1.28442099  0.800599143
#> 53  053  0.67192024  0.874801748
#> 54  054  1.16465283  0.794660294
#> 55  055  1.64872469  0.332549613
#> 56  056 -1.01255407 -0.741207303
#> 57  057  0.76918429  1.446393469
#> 58  058 -0.11989740  1.756598286
#> 59  059  0.30376451 -0.120892427
#> 60  060 -1.34970952 -0.687980575
#> 61  061 -0.30150763 -0.995837339
#> 62  062  1.60117433  0.555527179
#> 63  063  0.55762770 -1.217225320
#> 64  064 -0.16243980 -0.414908982
#> 65  065  0.09328992 -0.914007604
#> 66  066 -0.07398904  1.236829080
#> 67  067  0.37347859  0.357077593
#> 68  068  1.91696583  0.345217838
#> 69  069 -0.10201277 -0.387179721
#> 70  070 -0.88474380 -0.964218132
#> 71  071 -0.92175381 -1.532539192
#> 72  072 -0.69776492 -1.191633299
#> 73  073 -0.93301517 -0.719913152
#> 74  074 -0.10129275  0.385143672
#> 75  075 -1.42179515 -0.590262466
#> 76  076  0.46675051 -0.073917061
#> 77  077  0.72431242  0.215489323
#> 78  078  0.83280626  0.178325640
#> 79  079 -2.39708465  0.280989221
#> 80  080  0.05812231 -0.093386969
#> 81  081 -2.34626495 -1.314577019
#> 82  082  0.22575881 -1.521683528
#> 83  083  0.66222639 -1.053954093
#> 84  084 -0.87846676 -1.052327379
#> 85  085 -1.00175692 -0.500737782
#> 86  086  0.24036074  0.348509968
#> 87  087  1.67312230  0.406864056
#> 88  088  1.23422043  1.406654682
#> 89  089 -1.71255130 -1.034723721
#> 90  090  0.59481853  0.283191980
#> 91  091  1.41209633  0.413447423
#> 92  092 -0.03340899 -1.466782646
#> 93  093  1.00463782  1.882775969
#> 94  094 -0.94723398 -1.491221276
#> 95  095  1.61663959  0.534451594
#> 96  096 -0.76702195  0.908476994
#> 97  097 -0.97803979 -0.227897785
#> 98  098 -0.37209884 -1.799533789
#> 99  099  0.96719256  0.837103305
#> 100 100  0.19698870 -0.107680733

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.803268200  1.70406517
#> 2   002  0.369054156  0.21436669
#> 3   003 -0.002015923 -0.28780725
#> 4   004  0.626820733  0.46577935
#> 5   005 -0.828808086 -0.06821631
#> 6   006  0.593032832 -0.27867944
#> 7   007  0.740649521  0.11554293
#> 8   008 -0.692672491  0.25817566
#> 9   009  0.339379246 -1.51460793
#> 10  010 -0.981109822 -0.02677955
#> 11  011 -0.223240904 -0.39097120
#> 12  012 -0.151856373 -0.24833743
#> 13  013  0.094962451 -1.22013777
#> 14  014 -0.792866950 -0.43706965
#> 15  015 -1.346326251 -2.00752150
#> 16  016 -0.580420821 -1.19013822
#> 17  017 -2.026435901 -2.51013976
#> 18  018  0.116906375 -0.17365882
#> 19  019 -0.273624769 -1.09388214
#> 20  020 -1.250418757  0.64924136
#> 21  021 -0.080130894  0.42599118
#> 22  022 -1.410025013 -0.61826562
#> 23  023  1.017149443  1.45760225
#> 24  024 -0.316283779  0.01615637
#> 25  025 -2.561880582 -1.31293287
#> 26  026  1.007813742 -0.80292120
#> 27  027  0.223980559 -0.22458447
#> 28  028 -1.010093424 -0.01733187
#> 29  029 -0.056703101  0.94395758
#> 30  030 -0.231637790  0.01072217
#> 31  031 -1.149337988 -1.13567274
#> 32  032 -0.053569722  1.80248143
#> 33  033 -0.697518867 -0.49439854
#> 34  034 -0.570490003 -2.11013461
#> 35  035 -2.094875471 -0.27848358
#> 36  036 -1.731028268 -0.69082076
#> 37  037 -0.409069818  0.90312187
#> 38  038  1.112090220  1.06116498
#> 39  039  0.037175633 -0.36844986
#> 40  040  0.502106452  0.77254105
#> 41  041 -0.516114093 -0.33039561
#> 42  042 -1.534403914 -0.65046193
#> 43  043  1.710152912 -0.89772966
#> 44  044 -0.768440919 -1.00872967
#> 45  045  0.293052572  0.14051786
#> 46  046  0.153813356 -0.04006951
#> 47  047 -1.005069565 -0.75324042
#> 48  048 -0.102431107 -0.95239333
#> 49  049  0.754649757 -0.52248400
#> 50  050 -0.212334742  0.09409732
#> 51  051  0.142404657 -0.82507719
#> 52  052  0.805945952  0.27052235
#> 53  053  1.038845339  0.48981005
#> 54  054 -0.262120861 -0.76922330
#> 55  055  1.569929247  1.58801673
#> 56  056  0.660643850  1.61963541
#> 57  057  1.965028733  0.12799145
#> 58  058 -0.867768217  0.46346845
#> 59  059 -0.099227513  0.90205482
#> 60  060  0.954429387 -0.94738304
#> 61  061 -0.199012075  0.45041622
#> 62  062 -0.165458468  1.44583759
#> 63  063  1.280651727 -0.02290776
#> 64  064 -1.272525702 -0.67169851
#> 65  065 -0.937894317 -0.83988538
#> 66  066  1.374713713 -0.01242820
#> 67  067  0.806177660 -0.47565264
#> 68  068  1.807971421  0.52333474
#> 69  069 -0.964508206 -0.27076837
#> 70  070 -1.544341318 -1.55081136
#> 71  071 -0.122773864 -0.56186659
#> 72  072  0.145522691 -0.87873268
#> 73  073  0.018337944  0.31608297
#> 74  074  0.262753384  0.75117356
#> 75  075  0.174389726  1.33167516
#> 76  076 -0.280575159 -0.26804427
#> 77  077 -0.064302451 -1.35443135
#> 78  078 -0.762322359  1.12661566
#> 79  079 -0.668847307  0.61223565
#> 80  080  1.371067091  1.14000111
#> 81  081 -0.555156059 -0.50041293
#> 82  082 -0.362965151 -1.52003122
#> 83  083  2.150620675  0.47280803
#> 84  084  0.685939065  0.64694496
#> 85  085  0.999338185  0.95868700
#> 86  086 -0.283821623 -0.87054656
#> 87  087  0.536236945  0.96965642
#> 88  088  0.587366010  0.58682226
#> 89  089  0.992291626  0.75602302
#> 90  090  1.547746729  0.11699786
#> 91  091  0.054879775 -2.10766266
#> 92  092 -0.260114377 -0.06724657
#> 93  093 -0.781011880  0.03474587
#> 94  094  1.449263053  0.53547341
#> 95  095  0.173795751  1.34090466
#> 96  096  2.262757941  1.08434363
#> 97  097  3.069437111  1.82215183
#> 98  098 -1.681980741 -0.01779079
#> 99  099 -1.287971395 -0.62153995
#> 100 100 -0.795277621 -1.45764452