Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5567791
#> Y_2 0.5567791 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1       Y_2
#> 1 1.1498223 1.2184981
#> 2 0.4554497 0.2256333
#> 3 0.2972852 1.0644698
#> 4 1.5686708 1.3895510
#> 5 1.4340883 0.4049487

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.599827776  1.761516458
#> 2   002  2.418279911  2.033339973
#> 3   003 -1.119456521 -1.849989174
#> 4   004 -1.356264428  0.938582987
#> 5   005 -1.005838878 -0.876559540
#> 6   006 -0.494296213 -1.792839704
#> 7   007 -1.206771059 -0.151451674
#> 8   008  1.535075210  0.251895589
#> 9   009 -1.578078550  0.079001005
#> 10  010 -0.609433051  0.574109780
#> 11  011  1.964248355  1.400145978
#> 12  012  0.549956144 -0.228551228
#> 13  013  0.278294376 -0.559399885
#> 14  014 -0.828186819 -2.375340383
#> 15  015 -0.925024709 -0.060522883
#> 16  016  0.781763453  0.631784850
#> 17  017 -1.390751777  0.234597912
#> 18  018  1.115250302  0.532748457
#> 19  019 -0.279988901  0.025884014
#> 20  020 -1.954823662 -1.285747193
#> 21  021 -1.072380784 -0.701050972
#> 22  022  1.111253763 -0.384417030
#> 23  023 -0.798353331 -1.464905608
#> 24  024  1.238813941 -0.843383163
#> 25  025 -1.000718518 -0.396508707
#> 26  026  0.266644189 -0.149390240
#> 27  027  1.280130100  0.454795201
#> 28  028  1.723384616 -0.593865035
#> 29  029  0.748337240 -0.260821688
#> 30  030 -0.722439341 -1.745941150
#> 31  031 -0.085337870 -0.252789863
#> 32  032  0.863256307  0.214513245
#> 33  033  0.221777317 -0.044276832
#> 34  034  0.266095351  0.289680988
#> 35  035 -2.093159344  0.055378859
#> 36  036 -1.138098824  0.200245872
#> 37  037 -1.902540951 -0.802924734
#> 38  038 -0.002865017 -0.658984567
#> 39  039 -1.137229653 -0.512699527
#> 40  040 -1.399531128 -2.177734740
#> 41  041 -1.715431652 -0.459861991
#> 42  042 -1.014704170 -0.555994004
#> 43  043  0.244278336  0.852340161
#> 44  044 -0.180689877  1.231568848
#> 45  045 -2.246024452 -1.074155845
#> 46  046  0.595534226 -0.004579086
#> 47  047 -1.954051174 -0.321557301
#> 48  048 -1.154462280  0.389996278
#> 49  049  0.117502747  0.644186559
#> 50  050  0.032716587  0.334392016
#> 51  051 -0.111912934 -0.162714455
#> 52  052 -0.351615694 -0.180207535
#> 53  053  0.313592983  0.485924538
#> 54  054  0.211965675 -0.039183927
#> 55  055 -0.665553450  0.453051926
#> 56  056 -0.576497474  0.091468666
#> 57  057  0.872432048  0.334471953
#> 58  058 -0.397705087  0.156278165
#> 59  059  0.472133926  0.526572908
#> 60  060  0.560906399  0.120965613
#> 61  061  0.017229585  1.197081056
#> 62  062  1.464530306 -1.836996508
#> 63  063  0.435697389  1.806829073
#> 64  064 -0.553151807 -0.190824568
#> 65  065 -0.528130545 -2.569335706
#> 66  066 -0.910211926  0.177997795
#> 67  067  0.164779021  0.554994240
#> 68  068 -0.027044852  0.238995248
#> 69  069  0.613327177 -1.071089058
#> 70  070  0.400362059  0.924569159
#> 71  071  0.305035160 -0.364782721
#> 72  072 -1.047042367  0.761512664
#> 73  073 -1.369333972  1.613494737
#> 74  074 -1.307912161 -1.500770108
#> 75  075 -3.504153565 -0.707910493
#> 76  076  1.185073791  1.239190368
#> 77  077  0.711511769  2.604683350
#> 78  078  1.342006830  0.690433200
#> 79  079 -0.040562497 -0.590424371
#> 80  080  0.709206176  0.633337009
#> 81  081  0.522724683 -0.063337189
#> 82  082 -2.136729759 -1.005201978
#> 83  083  1.294725924  1.097501194
#> 84  084 -0.423722810 -0.419706215
#> 85  085 -1.708287268  0.400523993
#> 86  086  0.120662528 -0.114396476
#> 87  087 -1.479237466  0.088286446
#> 88  088 -0.736800036  0.357325232
#> 89  089 -1.309660683 -0.747670493
#> 90  090  0.946013490  0.206816086
#> 91  091 -0.801881548 -1.851158765
#> 92  092 -1.496196155 -0.217126482
#> 93  093 -0.850104299  0.100735039
#> 94  094 -0.261467327 -1.041394862
#> 95  095 -0.822411148  0.574770803
#> 96  096  0.185082261 -0.534924957
#> 97  097  0.234041194 -1.318897983
#> 98  098  1.041098044 -0.467809852
#> 99  099  1.314663119 -0.511109408
#> 100 100  1.643702962  1.473993774

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.526926873  0.01971681
#> 2   002 -0.871880357 -0.22314233
#> 3   003 -0.982948288 -0.15518996
#> 4   004  0.755663736  1.91695107
#> 5   005  0.221049081  0.51726448
#> 6   006 -1.181289710 -2.54608738
#> 7   007  0.048328874 -0.59071288
#> 8   008  1.336959475 -0.08286579
#> 9   009 -0.897579260 -0.54692090
#> 10  010 -0.587669020  0.94656789
#> 11  011 -0.485841056 -0.58222000
#> 12  012  0.569710099  1.97951455
#> 13  013  0.844689818 -0.09787025
#> 14  014 -0.349260287 -0.58993770
#> 15  015 -1.765129335 -0.30756768
#> 16  016 -0.126656389  1.96744444
#> 17  017  0.862144045 -0.60059079
#> 18  018 -0.336474951 -0.37805603
#> 19  019  0.222876339  0.22156707
#> 20  020 -0.784638791 -0.66926526
#> 21  021 -0.112608373 -1.34036773
#> 22  022 -1.639300981 -1.74448243
#> 23  023 -0.254211632  0.51823865
#> 24  024  1.253361174  0.58919777
#> 25  025  0.527475401 -0.06939985
#> 26  026  0.007029622 -0.19329767
#> 27  027  1.476931556  0.40337559
#> 28  028  0.418213019  2.51001158
#> 29  029  1.637102970  1.40874308
#> 30  030 -1.760019149 -1.79561001
#> 31  031  1.964203190  2.07922079
#> 32  032 -0.344564087 -0.75757823
#> 33  033  0.478876585  1.13772048
#> 34  034 -1.231919119 -1.54110536
#> 35  035 -0.482678569 -0.59499243
#> 36  036 -0.941228151 -0.29016671
#> 37  037  0.104584430  0.05093379
#> 38  038 -0.607393280 -0.57554843
#> 39  039  1.038154519  1.36159824
#> 40  040  0.667957361 -1.10301562
#> 41  041 -2.684708715 -1.92757291
#> 42  042 -0.451041270 -2.27065310
#> 43  043 -0.124502355 -0.71840511
#> 44  044  0.827045305 -0.02898323
#> 45  045 -0.529359505  0.64236000
#> 46  046 -0.343751789 -0.64627686
#> 47  047 -1.008531602  0.21986836
#> 48  048  0.763751954  1.00055596
#> 49  049  1.438787748 -0.37022666
#> 50  050  1.854906070  1.43480337
#> 51  051  0.206213077  0.22014601
#> 52  052 -0.447198648 -0.79692980
#> 53  053 -0.157285942  0.07200422
#> 54  054 -0.096598564 -0.69582653
#> 55  055  0.040381323  0.27916060
#> 56  056  0.355007959  1.54830704
#> 57  057 -1.660966765  0.39355985
#> 58  058 -0.152764212  0.34589053
#> 59  059 -1.009065569 -0.58980637
#> 60  060 -0.270388548  0.85248457
#> 61  061  0.011100932  1.65837712
#> 62  062  0.152346519  0.52514198
#> 63  063  2.359292416  2.28418591
#> 64  064  1.714519581  0.73640651
#> 65  065 -0.254280713  0.41696016
#> 66  066 -0.680791083  0.07358399
#> 67  067  0.452108129  0.32005650
#> 68  068  0.221053841 -0.10797619
#> 69  069  0.180569647  0.47975433
#> 70  070  1.497231033  1.07937453
#> 71  071 -0.383583918 -0.28568076
#> 72  072  0.361655881 -0.01292298
#> 73  073  0.795963048 -0.13742125
#> 74  074  0.812131323  0.09848541
#> 75  075  1.221459443 -0.23978684
#> 76  076  1.588415778  2.45124506
#> 77  077  1.651227845  1.56582783
#> 78  078 -0.365716450  0.09925878
#> 79  079 -0.573384837 -0.70632954
#> 80  080 -0.073171754 -0.23621235
#> 81  081  0.237621621  0.31747190
#> 82  082  0.143306451  0.91684264
#> 83  083 -0.293914958  1.07079022
#> 84  084  1.064415225  0.01246321
#> 85  085 -0.626819023 -1.13091800
#> 86  086  3.090336232  1.23298654
#> 87  087 -0.990720408 -0.92857807
#> 88  088 -1.665603204 -0.50319399
#> 89  089 -0.108371984 -1.18679742
#> 90  090  0.163478482 -0.44079884
#> 91  091 -0.906606153 -0.48529923
#> 92  092 -0.290740720  1.15845106
#> 93  093 -0.804069034  0.94097734
#> 94  094  0.756942282 -0.34316762
#> 95  095 -1.394493107 -0.93294115
#> 96  096  0.924927279  0.74819095
#> 97  097  0.787109341  0.06330591
#> 98  098  1.455269294 -0.61020149
#> 99  099  0.373403740 -0.11425936
#> 100 100 -2.548254178 -0.61513252