Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.540622
#> Y_2 0.540622 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1  0.943884737 -0.3311004
#> 2 -0.744979340 -0.1672756
#> 3  0.006384334  1.4411605
#> 4  1.157900030 -0.2509204
#> 5  1.568532896  0.3140199

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.06470547 -1.089250221
#> 2   002 -2.73387011 -0.712276134
#> 3   003 -1.46499943 -1.271353280
#> 4   004  0.10189368 -1.026412106
#> 5   005 -0.89498621 -1.056111625
#> 6   006  0.42952752  1.454322880
#> 7   007 -0.99322019 -0.388629155
#> 8   008  0.39602374  0.873397858
#> 9   009 -0.11262518 -0.952697748
#> 10  010 -1.31237160 -0.534752059
#> 11  011  0.71042194  0.550322772
#> 12  012 -0.22558018 -0.642958925
#> 13  013  1.32318251  0.719111937
#> 14  014  0.17130515  0.638297761
#> 15  015 -0.84122140  0.817693860
#> 16  016 -1.58137595 -1.167508037
#> 17  017  0.94049044 -0.460382621
#> 18  018 -0.31969710  1.377008630
#> 19  019  1.41718903  0.293414158
#> 20  020  0.41702750 -0.479675978
#> 21  021 -0.74051980  0.508501266
#> 22  022  0.18362787 -0.825996859
#> 23  023  0.90919548  1.126248731
#> 24  024 -1.02185537 -2.301387994
#> 25  025  0.14372640  0.793070786
#> 26  026 -0.18969178  0.866020533
#> 27  027  1.41464103  0.119062434
#> 28  028  0.69617972  0.522919562
#> 29  029  1.05503565  0.490003888
#> 30  030  0.27586971 -0.522767122
#> 31  031  0.96593156  2.632420604
#> 32  032 -0.21705952  0.131072496
#> 33  033  0.20601410 -1.074682290
#> 34  034  0.53983802  0.969925553
#> 35  035  0.48585544 -0.287892532
#> 36  036 -0.06867045  0.620703212
#> 37  037  1.43017866  0.258080663
#> 38  038 -0.21103574 -0.638066260
#> 39  039  0.82518145 -0.093380282
#> 40  040  0.33865130  0.497262758
#> 41  041  0.31951778 -0.817877178
#> 42  042 -0.62937249  0.391784085
#> 43  043 -1.05606699  0.249635984
#> 44  044  0.74156239  0.861457600
#> 45  045 -0.28081266  0.149486699
#> 46  046  1.05355450 -0.158921259
#> 47  047 -0.43434512 -0.404601933
#> 48  048  1.43738821 -0.021209258
#> 49  049 -0.57966221 -1.134159175
#> 50  050  0.51400409 -0.729200710
#> 51  051  0.13483751 -1.110091512
#> 52  052  0.46620101  1.093955054
#> 53  053 -0.48668129 -0.918515421
#> 54  054 -0.74559893  0.966029413
#> 55  055 -0.92429493  0.051094705
#> 56  056 -0.79704985  0.046623436
#> 57  057 -0.31108371  1.132732066
#> 58  058 -0.45382589  0.776972337
#> 59  059  2.02485032  0.911627046
#> 60  060 -1.30545580 -2.731428258
#> 61  061 -0.46161321  0.213172720
#> 62  062  1.44315598 -0.877922347
#> 63  063 -1.46025664 -0.534815148
#> 64  064 -0.22348579  0.847388387
#> 65  065  0.78496422  0.489427943
#> 66  066  1.08674870  2.061082204
#> 67  067 -0.65596244  0.248485626
#> 68  068 -1.57717089 -2.385495054
#> 69  069 -0.07626883  0.411599139
#> 70  070  0.28215618  1.466025830
#> 71  071 -1.40955646  0.164921089
#> 72  072  1.90781811  1.411881163
#> 73  073 -0.28630333 -0.891395178
#> 74  074 -1.52363873 -0.920580011
#> 75  075 -0.00375961  0.817131118
#> 76  076 -1.74500849 -2.297323038
#> 77  077 -1.19852557 -0.005036327
#> 78  078 -0.57212819  0.291783094
#> 79  079  0.98073920 -0.407804300
#> 80  080 -1.04263906 -0.506661678
#> 81  081  1.03097721  1.994002866
#> 82  082 -1.32335504 -1.260857410
#> 83  083  0.98767702  0.059610852
#> 84  084  0.43378862 -0.727140649
#> 85  085  1.43334010  1.340319574
#> 86  086 -0.66266416 -0.432126393
#> 87  087  2.24332270  0.782918184
#> 88  088  0.14880447  0.452348502
#> 89  089 -0.40537281 -1.102151466
#> 90  090 -0.79105237  1.077233233
#> 91  091 -0.30937722  0.544924704
#> 92  092 -0.70275966  0.288297878
#> 93  093 -0.26053023  0.285567255
#> 94  094  0.11681832 -0.087277543
#> 95  095  0.43016550  0.251524361
#> 96  096 -0.32063479  0.920272595
#> 97  097  0.36088126  0.767929309
#> 98  098 -0.62435608 -0.673141329
#> 99  099  1.45707045  0.553775432
#> 100 100 -0.12678478 -0.199474341

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.274292700 -0.25883289
#> 2   002  2.321574770  0.40954606
#> 3   003 -1.976267574 -1.43418997
#> 4   004 -1.019927484 -0.66400432
#> 5   005 -1.523064148  0.36329488
#> 6   006 -0.903248236  2.03535625
#> 7   007 -0.985051571 -0.71861067
#> 8   008 -0.367530369 -0.39095388
#> 9   009  0.770627039  1.21205022
#> 10  010  1.052497418  0.29539705
#> 11  011  1.048293938 -0.12664633
#> 12  012 -0.423302951 -0.94136339
#> 13  013  0.358522045 -2.18104876
#> 14  014  0.272462573  0.15449526
#> 15  015  1.871229891  1.34662441
#> 16  016 -0.616608690  0.63282147
#> 17  017  1.499172308  2.42430086
#> 18  018  0.698700362  0.31735311
#> 19  019 -1.261221581 -0.62490180
#> 20  020  1.508331780  0.30570615
#> 21  021  0.705476328  1.32417804
#> 22  022 -0.351827006  0.18756884
#> 23  023  1.013805849 -0.50278671
#> 24  024  1.150620226  1.06486994
#> 25  025 -0.129227068  0.89428314
#> 26  026  0.175645649 -0.27295454
#> 27  027  1.710136988  1.95696618
#> 28  028  0.181158652  0.51915559
#> 29  029 -0.596142264 -0.66879350
#> 30  030  2.706175656  1.76856605
#> 31  031  1.131475052  0.76859683
#> 32  032  1.264187791  0.46452409
#> 33  033 -0.246479483  0.99916263
#> 34  034  0.185672727 -0.51129074
#> 35  035 -2.247641583 -0.67318578
#> 36  036 -0.042029966  0.32662124
#> 37  037 -1.134127534 -1.21891438
#> 38  038 -0.405170756 -0.39796858
#> 39  039  0.119907349 -0.34523449
#> 40  040  1.010654100  0.76897004
#> 41  041 -1.230295691  1.42520796
#> 42  042 -0.235712627  1.37721102
#> 43  043  0.038283816  0.47179460
#> 44  044 -2.142012695 -2.52726573
#> 45  045  0.327033989  0.53418318
#> 46  046  1.770495947  0.65359389
#> 47  047 -0.640107288 -0.37440663
#> 48  048  0.499574170  0.14115279
#> 49  049 -0.920094574 -0.77057034
#> 50  050 -0.370928015 -0.35410270
#> 51  051 -1.111799947 -2.09187310
#> 52  052 -0.152381383 -1.77689227
#> 53  053  0.566075581  0.05803451
#> 54  054  0.035620251  0.79387349
#> 55  055 -0.297473862  0.88947576
#> 56  056 -0.473424149 -1.57434720
#> 57  057 -0.494295115 -0.57769437
#> 58  058 -1.426187626 -0.98362564
#> 59  059 -0.789957466  1.51820694
#> 60  060 -1.019659508  0.20249357
#> 61  061  0.871430825  1.37914649
#> 62  062  1.522047548  2.37969572
#> 63  063 -0.573876663 -0.54773154
#> 64  064 -0.679932304  0.10627618
#> 65  065  0.353791506  0.70901399
#> 66  066  0.735767273  0.78401476
#> 67  067  2.014629711  2.33763746
#> 68  068  1.043317919  0.62459126
#> 69  069  1.412481944  2.50684138
#> 70  070  0.011432912  1.79526347
#> 71  071 -0.356298069 -1.72051686
#> 72  072 -0.663508420 -1.33939292
#> 73  073 -0.207225210  1.19613441
#> 74  074 -0.008129769 -0.78757852
#> 75  075  1.291429276  1.32389881
#> 76  076 -0.835618082 -0.68754067
#> 77  077  0.506217763 -0.38890065
#> 78  078 -0.920445161  1.21889167
#> 79  079 -1.129181911 -0.47951927
#> 80  080  0.616156081 -0.92199759
#> 81  081  1.088571944  0.53878998
#> 82  082 -0.328015384 -1.43492599
#> 83  083 -2.110245299 -2.39357639
#> 84  084  1.064536841  0.38322235
#> 85  085 -0.289817210 -0.38056086
#> 86  086 -1.423846153 -0.45976656
#> 87  087  2.366273685  0.74003523
#> 88  088  1.191120335  0.27359424
#> 89  089 -0.153925644 -0.19756909
#> 90  090 -1.055253083 -2.09271023
#> 91  091 -0.258622605 -1.37025510
#> 92  092  2.684737273  2.55641910
#> 93  093  0.799588535  0.19492294
#> 94  094 -0.085005753  0.18009689
#> 95  095  0.871043787  0.67345627
#> 96  096 -1.318221016 -2.00832991
#> 97  097 -0.445166004 -1.43858901
#> 98  098  0.669627332 -0.30465750
#> 99  099  2.501473941  0.90935544
#> 100 100 -0.289592290 -0.46418824