Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5094747
#> Y_2 0.5094747 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1 -0.543446975  0.2753132
#> 2 -0.606599394 -0.7828541
#> 3  0.207013343  1.6251791
#> 4 -0.671050279 -1.5946353
#> 5 -0.009411843  0.6131841

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1           Y_2
#> 1   001  1.604963172 -1.0018397137
#> 2   002  0.250173336 -1.4784547792
#> 3   003 -2.423584883 -2.8623211216
#> 4   004 -0.680383675 -0.0418863225
#> 5   005  0.224010486  0.0241907967
#> 6   006  0.825082641  1.1816532636
#> 7   007 -0.227851463 -0.3943789217
#> 8   008 -0.637902708 -0.1804196030
#> 9   009 -1.352713214 -2.6214771797
#> 10  010 -0.754033950  1.7950344162
#> 11  011  0.035240810  0.8292618025
#> 12  012 -0.629320442 -0.4641810762
#> 13  013 -1.475000427 -1.7241294540
#> 14  014 -0.303949221 -0.4785141886
#> 15  015  0.514507420 -0.5186385900
#> 16  016 -0.404998296 -0.3936173656
#> 17  017 -1.691026196 -2.6084241697
#> 18  018  0.225562633 -0.3025015228
#> 19  019  0.616404272  0.1239312279
#> 20  020  1.123114005 -0.6706708883
#> 21  021 -0.352172069 -0.0749281430
#> 22  022 -1.401854691 -1.0275312368
#> 23  023  1.955529409  2.3341559705
#> 24  024  0.771017802  0.8833164808
#> 25  025 -0.482314032 -0.1420020024
#> 26  026  0.363777927  1.1781847953
#> 27  027  0.252200104  0.5737675833
#> 28  028  0.071214655 -1.1218021616
#> 29  029  0.092988831  0.6608378944
#> 30  030 -0.403899908 -0.5019387193
#> 31  031  0.206086353  0.5910127003
#> 32  032  0.059366275 -0.3018086574
#> 33  033  0.532186293  0.0168086391
#> 34  034  0.234300835 -0.3248320997
#> 35  035 -0.145813049 -0.3628614384
#> 36  036 -0.249829098 -0.3053317213
#> 37  037  0.163211390  1.5468137347
#> 38  038  1.121403044  0.4315116666
#> 39  039 -0.675021972 -0.6933730849
#> 40  040 -0.695851532  0.0950156664
#> 41  041  0.510074675  1.0863424502
#> 42  042  0.576740585  1.2973438370
#> 43  043 -0.255314047 -0.1442022750
#> 44  044 -0.504909904 -0.1030991875
#> 45  045  1.443149849  1.7690372159
#> 46  046  0.645527952  0.4740748301
#> 47  047  0.426724777 -0.4784701915
#> 48  048 -0.911401218 -0.4180663056
#> 49  049  1.792194632  1.9820157890
#> 50  050 -2.359120398 -1.8425162209
#> 51  051 -0.242075609 -1.3425030157
#> 52  052  0.012721097  1.3678974583
#> 53  053  0.578738499 -0.3645279083
#> 54  054  1.986051035  3.1082660004
#> 55  055  0.063394324 -1.5457876727
#> 56  056  0.002107133  1.9369755557
#> 57  057  0.882144105  0.7622023524
#> 58  058 -0.770816729 -0.6133811082
#> 59  059  0.326949938  0.1269405946
#> 60  060  0.332468883 -0.0002970183
#> 61  061  0.170405596  0.4856171264
#> 62  062  1.165746377  0.8516040159
#> 63  063 -0.090713837 -1.5028999721
#> 64  064  1.883435443  0.4451248737
#> 65  065  0.153054321 -0.3182229658
#> 66  066  0.843459237 -2.1177164645
#> 67  067 -0.598398857  1.1946928979
#> 68  068 -0.303586381 -0.8119672049
#> 69  069 -0.161716759 -2.3203329625
#> 70  070 -1.077384007 -0.6603857130
#> 71  071 -0.870690141 -0.0646870128
#> 72  072  0.291361263 -0.7000707465
#> 73  073  0.141966798  0.2074776683
#> 74  074  1.035146999 -0.1913533101
#> 75  075 -0.542422506  1.4112320178
#> 76  076 -0.659093205 -0.6042712316
#> 77  077 -0.212277842  0.1661747261
#> 78  078 -1.173267969  0.0413165662
#> 79  079  2.182553922  0.8628126830
#> 80  080 -0.757355646  0.7525549673
#> 81  081  1.436882878 -0.4016295225
#> 82  082 -0.052573694 -0.5809013891
#> 83  083  1.017129507 -0.8628760831
#> 84  084  0.682412847  0.3163501292
#> 85  085 -0.828712417 -0.1677833821
#> 86  086 -1.576940252  0.4833381294
#> 87  087  0.045055112  2.1917646610
#> 88  088  0.117428439  0.2032482148
#> 89  089 -0.861969364  0.8487722188
#> 90  090 -0.247364595  0.4992766979
#> 91  091 -0.762228978 -1.8210023707
#> 92  092 -0.792449687 -0.4951128774
#> 93  093 -1.338217043 -0.5145663048
#> 94  094  0.953369598  0.6065891213
#> 95  095  1.859113146  1.6111376590
#> 96  096 -0.674300078 -0.1300542293
#> 97  097  0.527638438  1.2986504056
#> 98  098 -0.578654729 -0.1297005114
#> 99  099  0.302674587 -2.3386465076
#> 100 100 -1.290897551 -0.3029993461

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.510288799 -1.27968586
#> 2   002  1.124197556  1.64500430
#> 3   003  0.379516060 -1.17815175
#> 4   004 -0.460056897  0.04212745
#> 5   005  0.831380774  2.31995541
#> 6   006 -1.299543068 -0.19128815
#> 7   007 -0.874616505  0.33755265
#> 8   008  1.250732655  0.98878886
#> 9   009 -1.176404909  1.33672875
#> 10  010 -1.582637852 -0.44961322
#> 11  011  2.284896951  1.93162982
#> 12  012 -0.075105175 -0.82035912
#> 13  013 -1.648671113 -1.87119043
#> 14  014  0.285767726  1.08537282
#> 15  015 -0.633340682  0.42832655
#> 16  016 -0.289232647 -0.35261879
#> 17  017  0.656384158 -0.16015951
#> 18  018  1.669411333  2.17071029
#> 19  019  1.538176706  0.09636014
#> 20  020 -0.856851344  0.67066231
#> 21  021  0.366351038  0.24423120
#> 22  022  1.373979658  0.50769835
#> 23  023  1.011915965  0.26835043
#> 24  024  0.992652353  1.91542002
#> 25  025  0.270133584  0.93507384
#> 26  026 -0.158256514 -1.32791930
#> 27  027 -0.692172709  0.10945229
#> 28  028  0.176351404 -0.61436631
#> 29  029  1.626588396  1.09084802
#> 30  030 -0.094471780  0.24065411
#> 31  031 -1.638440550 -1.24274548
#> 32  032 -0.488611943 -1.28167608
#> 33  033 -0.509339249 -0.98148854
#> 34  034 -1.379180031 -0.65517362
#> 35  035 -0.757915987 -0.66859569
#> 36  036 -0.382847554 -0.64188663
#> 37  037 -1.163937649 -1.36634031
#> 38  038 -0.159973220 -0.24692434
#> 39  039  0.267472822 -0.76449189
#> 40  040  0.171479290 -0.46339218
#> 41  041 -0.271325148 -0.21771504
#> 42  042  0.466989798 -1.18087507
#> 43  043  0.778794747 -1.78105621
#> 44  044  1.009723741 -0.75789436
#> 45  045 -0.399647545  0.05796671
#> 46  046 -1.115774000 -2.26358630
#> 47  047  2.546994772  0.66591043
#> 48  048  0.828215697  0.53660342
#> 49  049 -1.653704180  0.26191919
#> 50  050 -1.192711077 -1.48126122
#> 51  051 -0.116741330  0.70321442
#> 52  052 -0.612560647  1.33616583
#> 53  053  0.678711360  1.16784203
#> 54  054 -1.229026268 -0.77517461
#> 55  055 -0.705519827 -0.82573912
#> 56  056  0.632720628  1.48566237
#> 57  057  0.402520323 -1.05853054
#> 58  058 -0.592328901 -0.74440695
#> 59  059 -0.611950580 -0.29020606
#> 60  060 -1.220330322 -0.07413969
#> 61  061 -0.919455287 -1.83581701
#> 62  062  0.234948268 -0.27658962
#> 63  063 -0.265974905 -0.19437901
#> 64  064  0.002928402  0.87594797
#> 65  065  1.146414295  1.77941434
#> 66  066  1.423699226  1.48952698
#> 67  067  0.262442174  0.70643288
#> 68  068 -0.289565501 -0.68732352
#> 69  069  0.947360816  1.70990288
#> 70  070  0.534750654  1.08414545
#> 71  071 -0.700404641 -0.06432521
#> 72  072 -0.587476547 -0.48611491
#> 73  073  0.762015654  0.60301215
#> 74  074  0.780357532  0.35667166
#> 75  075 -0.412152070  0.09619573
#> 76  076 -0.152767375 -0.04930075
#> 77  077 -0.494496650 -0.77589213
#> 78  078  2.830890524  1.32143710
#> 79  079  0.018820893 -0.08333737
#> 80  080  1.647309873  2.36109019
#> 81  081  0.565835009 -0.70465130
#> 82  082  0.095890703  0.82211559
#> 83  083  0.591879090  0.14134228
#> 84  084  0.628234165  1.27569451
#> 85  085 -1.982035179 -1.00036642
#> 86  086 -0.135393995  0.10979097
#> 87  087 -1.659325666 -1.84735805
#> 88  088 -0.081719561  0.15702333
#> 89  089 -0.630063319 -1.39110737
#> 90  090  0.482321247 -0.68836268
#> 91  091 -0.134393007  0.43014801
#> 92  092  1.091463528  0.72850283
#> 93  093  0.115611130  1.15375800
#> 94  094 -0.032188424  1.12241953
#> 95  095  0.794357596  2.71002026
#> 96  096 -0.610482501  0.26802670
#> 97  097  0.381974637 -0.39426305
#> 98  098 -0.534652580 -1.34278915
#> 99  099 -0.831536611  0.82775109
#> 100 100 -1.133417169 -0.83794232