Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4742187
#> Y_2 0.4742187 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.7660036  0.3062674
#> 2  2.4038247  0.8066831
#> 3 -0.9424081 -1.1753379
#> 4 -1.7627936 -1.2394175
#> 5 -0.4310420  1.0705949

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.90673580 -1.017648753
#> 2   002 -0.22699126 -1.599955255
#> 3   003 -0.55631781 -0.022005020
#> 4   004 -0.41561340  0.661296280
#> 5   005 -0.14320885 -1.724017529
#> 6   006  0.38218718  0.071537390
#> 7   007  0.47714410  0.618730570
#> 8   008 -0.38369709 -0.734874996
#> 9   009 -1.01022073 -1.597333833
#> 10  010  0.32059850  0.201045414
#> 11  011  0.84181471  0.984818063
#> 12  012  1.06287458  0.318538073
#> 13  013  1.05076734 -0.318586535
#> 14  014 -0.22863098  0.085951788
#> 15  015  0.04842355 -0.375988915
#> 16  016  0.87059504 -0.028875132
#> 17  017 -0.59190348  0.191610320
#> 18  018 -0.61164813  0.554372632
#> 19  019  0.19599250 -0.281478566
#> 20  020 -0.69359808 -0.286073473
#> 21  021 -0.86973763  0.461457030
#> 22  022  0.26800913  0.873949955
#> 23  023  0.53794278  1.723870438
#> 24  024 -0.47679865 -0.752557190
#> 25  025  0.39911184  1.403250780
#> 26  026 -1.91150234 -1.497334025
#> 27  027  1.08038858 -0.207482946
#> 28  028 -2.19399823 -0.704481570
#> 29  029 -1.08029883 -1.759542920
#> 30  030  1.01367122 -0.082057189
#> 31  031 -1.35102751 -0.667468800
#> 32  032  1.88665037  1.066007724
#> 33  033  1.52085061  1.808269968
#> 34  034 -0.06549891 -1.324434305
#> 35  035  0.94612228 -0.361397729
#> 36  036 -1.91509874 -0.733326724
#> 37  037 -1.06859366 -1.005774916
#> 38  038  0.13039878  0.848445395
#> 39  039  0.82848345 -0.977073806
#> 40  040 -0.12477869  1.541381174
#> 41  041 -0.11237667  0.568102833
#> 42  042 -0.30077132  0.179909540
#> 43  043  0.03050228 -0.182337422
#> 44  044  0.94736467 -0.228617535
#> 45  045 -1.84170822 -0.071471776
#> 46  046  0.10861163  0.052529966
#> 47  047  0.76607381 -1.694446700
#> 48  048  0.36009806  1.745614886
#> 49  049 -0.44016875  0.079565035
#> 50  050 -0.47391211 -0.920741694
#> 51  051  0.51210074  1.165732917
#> 52  052  0.09966438  0.529284135
#> 53  053 -1.22694358 -0.381588076
#> 54  054 -0.05049205  0.246557613
#> 55  055 -0.97878476 -0.918151818
#> 56  056  0.41906068  0.123134533
#> 57  057 -0.47040944  1.129909601
#> 58  058 -0.04642853 -0.201766083
#> 59  059  0.58701126  0.724002596
#> 60  060 -1.29291836 -0.019569805
#> 61  061  1.59734852  0.841778101
#> 62  062 -1.39830778 -1.135886204
#> 63  063 -0.91803780 -0.408529424
#> 64  064  2.22654807  0.372829267
#> 65  065  1.14048860  0.388719429
#> 66  066 -1.13782004 -1.826841135
#> 67  067 -1.32094146 -1.446297651
#> 68  068  0.27236820 -0.220218753
#> 69  069  0.38193912  0.668549298
#> 70  070 -0.37721467  0.148026141
#> 71  071 -0.69458662  0.052816569
#> 72  072 -0.79613431  0.001525429
#> 73  073 -0.16906664 -0.131210463
#> 74  074 -0.95060761 -0.687117302
#> 75  075  0.22416082  0.112562678
#> 76  076  0.29010330 -0.257011563
#> 77  077  0.52593872  0.844111904
#> 78  078  1.11877075  0.445085720
#> 79  079  1.33785373  1.068597815
#> 80  080  1.14941679  0.784505966
#> 81  081  0.37707269 -1.853230392
#> 82  082  1.25808528  0.781990586
#> 83  083 -1.11337034 -0.127531258
#> 84  084 -0.32437033  1.395328853
#> 85  085  0.06594291 -0.035698362
#> 86  086 -0.34412829 -1.024320326
#> 87  087 -1.76142065 -1.132228848
#> 88  088  0.03631932  0.268547435
#> 89  089  0.07484981  0.154440523
#> 90  090 -1.40610194  0.020547496
#> 91  091  0.01199046  1.707506981
#> 92  092 -0.25653724 -0.842042915
#> 93  093  0.43193740 -0.714507396
#> 94  094  0.27004647  0.540589903
#> 95  095  0.27408216  0.256194934
#> 96  096  0.12871732  0.020760631
#> 97  097  0.04000746  0.626373824
#> 98  098  0.23938135  0.604107559
#> 99  099 -2.02868314 -1.016092416
#> 100 100 -0.65670799 -0.915121440

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.739440080  0.015049653
#> 2   002  0.019885241 -1.262337037
#> 3   003  1.006428137  0.449369035
#> 4   004 -0.923446506 -0.227650280
#> 5   005  0.611784823  2.389292672
#> 6   006 -0.085996891  0.004796683
#> 7   007  0.089802744  0.589400349
#> 8   008  0.122385363  0.377263538
#> 9   009 -1.150662357 -0.992964576
#> 10  010  2.589844759  1.980383702
#> 11  011 -0.174431353  0.747373876
#> 12  012 -0.423108088  0.158901967
#> 13  013 -0.501883119 -0.069691891
#> 14  014 -1.002808716 -1.690056288
#> 15  015 -0.485578324 -1.088844384
#> 16  016  0.239415076 -0.373974692
#> 17  017 -0.068320536  0.453426009
#> 18  018 -0.267953636 -1.799326261
#> 19  019  0.766283102  0.531906112
#> 20  020  3.265868226  2.678649104
#> 21  021  1.313817534  0.540866944
#> 22  022 -0.096401580  0.159475522
#> 23  023 -0.682328660 -0.872705133
#> 24  024  0.843132898 -0.706928022
#> 25  025  0.174299742 -0.395346155
#> 26  026  0.621571264  1.033382426
#> 27  027  0.111023028  2.277710127
#> 28  028 -0.417651664  0.467519101
#> 29  029 -0.904571874 -0.825455067
#> 30  030 -0.435736081 -1.137847761
#> 31  031 -1.813523205 -1.717672349
#> 32  032 -0.311277415  0.199934888
#> 33  033  0.266618858  0.789591794
#> 34  034  0.026781921  0.271517662
#> 35  035  2.616982296  1.861280009
#> 36  036 -1.363320099 -1.063975601
#> 37  037  1.542594726  0.007700294
#> 38  038 -1.314506109 -0.071219420
#> 39  039  0.639601433 -0.318985063
#> 40  040  0.084695227  0.644651560
#> 41  041 -0.253652550 -0.186032766
#> 42  042 -2.202614633 -0.758545147
#> 43  043 -0.643378948 -0.822148811
#> 44  044  0.383539258 -0.318669287
#> 45  045  0.282298393 -0.572311390
#> 46  046  2.677572129  0.602316870
#> 47  047 -2.248893357 -0.331192911
#> 48  048  0.560153404 -0.002450374
#> 49  049 -0.116811762  0.326560726
#> 50  050 -1.958290971  0.228226103
#> 51  051 -0.443430645  1.054486665
#> 52  052 -0.316040099 -1.675933516
#> 53  053 -1.414218420 -1.449094512
#> 54  054  1.308545837 -0.706334824
#> 55  055  0.279542623  0.190091409
#> 56  056  1.393131849  1.362052090
#> 57  057 -1.698511070 -1.865785592
#> 58  058  1.986850277  1.134148235
#> 59  059  0.283530300  0.664634992
#> 60  060  1.289570913  1.048683902
#> 61  061 -0.704193893 -0.031536202
#> 62  062 -0.426210625  0.414738952
#> 63  063 -0.026891776  0.734291696
#> 64  064  0.941863488 -0.055108616
#> 65  065  1.181875533  0.182887994
#> 66  066  0.439342748 -0.805450251
#> 67  067 -0.771497386  0.621441306
#> 68  068 -1.666633739 -0.231057632
#> 69  069  0.038045986  0.717335575
#> 70  070 -0.060151256 -0.168803686
#> 71  071  0.659346068  0.617931389
#> 72  072  0.526692076  0.074618023
#> 73  073  0.297448452  0.380241407
#> 74  074  1.004955392  0.192477030
#> 75  075  0.564823150  0.307773094
#> 76  076 -0.008422939 -0.182585159
#> 77  077  1.142982328  1.601573458
#> 78  078  0.018153505 -0.198900367
#> 79  079  0.638666033 -0.594364565
#> 80  080 -1.653129604 -1.006334160
#> 81  081 -0.886121553 -0.350741717
#> 82  082  2.126143605  1.792257111
#> 83  083  1.042424984 -1.409497116
#> 84  084  1.439101045 -0.512627149
#> 85  085 -0.967959955 -1.328005046
#> 86  086 -0.351076259 -1.295668172
#> 87  087 -1.316978942 -0.276259472
#> 88  088  0.042769455  0.329135209
#> 89  089  0.176712729  1.004031745
#> 90  090  1.164265860  0.331889293
#> 91  091  0.846785430  0.825029090
#> 92  092 -0.102525391 -0.268032669
#> 93  093  0.683864098  0.374024614
#> 94  094 -0.577602308 -0.411680809
#> 95  095 -1.278728151 -1.287513786
#> 96  096 -0.654110077 -0.948995080
#> 97  097  0.525525597 -0.097909683
#> 98  098  0.409354782  0.934219130
#> 99  099  0.334404796 -0.534576408
#> 100 100 -0.878009671  1.197909189