Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5104395
#> Y_2 0.5104395 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.5530659 -0.29050780
#> 2  1.6026292 -0.59895982
#> 3 -0.2322792 -0.03685247
#> 4 -0.3913953  0.13380887
#> 5 -0.2029793 -0.24021778

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.99444539 -1.05457339
#> 2   002 -0.38495655 -0.32823695
#> 3   003 -0.79704234  0.04497403
#> 4   004  0.48046057  1.80937986
#> 5   005  1.17582157 -0.35158169
#> 6   006  0.47972744  0.06621407
#> 7   007 -0.11622452  0.64596932
#> 8   008  1.46850511  0.14423126
#> 9   009  0.28179972  0.25492271
#> 10  010 -0.18379154 -0.54297441
#> 11  011  0.97609020  0.86657295
#> 12  012 -0.14922934  0.79107731
#> 13  013  0.25018062 -0.33607725
#> 14  014 -0.54782827 -0.61253714
#> 15  015  0.64356009  0.84617985
#> 16  016 -0.06117127 -0.73716223
#> 17  017  0.54259696  1.03800431
#> 18  018 -0.32530327  0.30270123
#> 19  019  0.68342368  0.39022887
#> 20  020  0.65894734 -0.91788260
#> 21  021 -0.23510151 -0.03250595
#> 22  022 -1.22197696 -0.35911611
#> 23  023 -0.54669051 -0.75817991
#> 24  024 -1.82651022 -1.23258265
#> 25  025 -1.10706131  1.18701581
#> 26  026  1.54083868 -0.91816581
#> 27  027  0.54154751  0.19884806
#> 28  028  0.93843965 -0.71125437
#> 29  029 -1.21499491  1.12705710
#> 30  030 -0.97976098 -2.15073163
#> 31  031 -1.39072792 -0.76373236
#> 32  032 -1.70151559 -1.28679354
#> 33  033  0.01491207 -0.02925026
#> 34  034  1.83618561  1.52994079
#> 35  035  0.61318789 -0.32277868
#> 36  036  0.17975187  2.01961447
#> 37  037  1.08533571  0.56677073
#> 38  038  0.71243962  2.15806620
#> 39  039  0.48374122  1.19770290
#> 40  040 -0.80418442 -0.40865565
#> 41  041 -0.50601868  0.02920719
#> 42  042  0.33753066  0.55184893
#> 43  043  0.40462560  0.96678685
#> 44  044 -0.26670189  0.33606554
#> 45  045 -0.72874416 -0.06297935
#> 46  046  0.09659773  0.45219323
#> 47  047 -0.54758035  0.36446707
#> 48  048 -0.07141255 -1.42869126
#> 49  049  0.87032803  0.40347802
#> 50  050  0.11131362 -0.48925652
#> 51  051  0.63413713 -0.14293227
#> 52  052 -0.14894949  0.36372521
#> 53  053  0.12239075 -0.43854137
#> 54  054  1.59144256  0.11917395
#> 55  055  0.41342052 -0.10457033
#> 56  056  1.12130227 -0.61241172
#> 57  057 -0.32195289  0.51632628
#> 58  058  0.04089439 -0.46144927
#> 59  059  1.01150085  0.36534292
#> 60  060 -0.11318351  0.48226423
#> 61  061 -0.28862283  0.31817744
#> 62  062  1.76445970  3.01523057
#> 63  063  0.83731488  0.38038147
#> 64  064 -0.32694566 -0.96297191
#> 65  065 -0.06813319  0.12168236
#> 66  066  1.71915221  1.05101600
#> 67  067 -1.54889123  1.38006656
#> 68  068  2.30115746  1.28485659
#> 69  069  0.58080423  1.16096862
#> 70  070 -1.16061852 -0.05510950
#> 71  071 -0.65180373 -1.15085441
#> 72  072  0.91466486  0.44984521
#> 73  073  1.23172029 -1.01443727
#> 74  074  0.56875441  0.14580034
#> 75  075 -1.17919527 -0.58428159
#> 76  076  1.31935837 -0.18609593
#> 77  077 -0.97614760 -0.87319556
#> 78  078  3.52013808  0.89465299
#> 79  079 -1.26089089 -1.20244435
#> 80  080  0.40279551 -0.02404819
#> 81  081  1.03364184 -0.13922041
#> 82  082 -2.01492927 -0.96640161
#> 83  083  0.55217289 -0.17222435
#> 84  084 -1.57856188 -2.29610989
#> 85  085 -2.74354310 -1.95811857
#> 86  086 -0.22088933  0.29613225
#> 87  087 -0.65242071  0.58407829
#> 88  088  0.72161469  1.11026647
#> 89  089  0.58236478 -0.04444019
#> 90  090 -0.32071637 -1.11130058
#> 91  091  0.77467515  0.09541497
#> 92  092  0.59323187  1.23234776
#> 93  093  0.37562165 -0.09125331
#> 94  094  0.71263967 -0.32626510
#> 95  095  0.68491947  0.10200697
#> 96  096  1.79810385  1.11842645
#> 97  097 -0.46951314  1.44001977
#> 98  098  0.19016434  2.13324215
#> 99  099 -0.13244024  1.22681289
#> 100 100  0.91852575  1.25900558

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.29295365 -0.807894041
#> 2   002 -1.31187502 -0.079657257
#> 3   003 -0.39439819 -0.650527525
#> 4   004  1.05374290  0.088371316
#> 5   005 -0.34435433 -0.415013433
#> 6   006  1.00393296  2.508374538
#> 7   007 -0.51924301 -0.921965440
#> 8   008 -0.26192443  1.101409394
#> 9   009  1.60796819  1.962754477
#> 10  010 -0.41525713  0.922351450
#> 11  011 -1.48028609 -0.433655690
#> 12  012  0.48093189  1.385898161
#> 13  013  0.34679390 -0.334249173
#> 14  014  1.72709076  2.994066219
#> 15  015  0.52432088 -0.040815101
#> 16  016 -0.02448599  0.577454297
#> 17  017 -1.85311653 -0.756611949
#> 18  018 -0.13821718 -1.176824287
#> 19  019 -0.32874566 -1.278102865
#> 20  020  0.66957928 -0.357197162
#> 21  021  0.87313185 -0.066005961
#> 22  022 -0.58723887 -1.185770483
#> 23  023 -0.19756542 -1.268446358
#> 24  024 -0.86942306  0.314791803
#> 25  025 -0.48597035 -0.942672438
#> 26  026 -0.42460577  0.119461143
#> 27  027 -1.69366101 -1.056324174
#> 28  028 -0.34027656 -1.305354777
#> 29  029 -0.38281765  0.324859301
#> 30  030  1.96996638  1.784327974
#> 31  031 -2.83338309 -2.656293996
#> 32  032 -0.41162851  0.292031345
#> 33  033  1.26359799  2.065715467
#> 34  034 -1.50652177 -0.140387061
#> 35  035  0.02699639  1.389083592
#> 36  036  0.40801606  1.282380613
#> 37  037  0.82696087  0.069730710
#> 38  038  0.11173103  0.226547551
#> 39  039 -1.18374044 -1.628199326
#> 40  040 -0.10647656  1.069803272
#> 41  041  0.43989951 -0.199488028
#> 42  042 -0.92606330 -0.267646657
#> 43  043 -0.35900947  0.789457901
#> 44  044 -0.52916035 -2.024939319
#> 45  045  0.81020726  0.155079093
#> 46  046 -1.15597745 -0.276036359
#> 47  047 -0.46528599  1.095566287
#> 48  048  0.85940777  1.666857050
#> 49  049 -0.56392111 -0.089791111
#> 50  050  1.09438511  0.332928757
#> 51  051  0.12980914 -1.856596774
#> 52  052 -0.36282882  0.071197531
#> 53  053  0.80185727 -0.340055263
#> 54  054  1.09141529  0.498054258
#> 55  055  0.23923311 -0.590878455
#> 56  056 -1.74658777 -0.157177007
#> 57  057  0.99922155  1.863371656
#> 58  058 -0.73053789 -0.875174483
#> 59  059  1.57339575  0.622339534
#> 60  060 -0.28421948  1.388434254
#> 61  061 -0.80935732  0.126110814
#> 62  062  0.24811923 -1.201873846
#> 63  063 -1.14328539 -0.837857773
#> 64  064 -0.20343267  0.479520216
#> 65  065 -0.95985937 -1.344062179
#> 66  066  1.36903489  1.901848766
#> 67  067 -1.67183649 -0.324167322
#> 68  068 -0.03454669  0.006903551
#> 69  069  0.86077838  1.985624192
#> 70  070 -0.52873799  1.335867824
#> 71  071  1.86917663  1.040777163
#> 72  072  0.15089014  1.054791185
#> 73  073 -0.06181080  1.396481906
#> 74  074  0.60581881  1.225132379
#> 75  075 -0.09393837  1.056533398
#> 76  076  0.63266856 -0.968716122
#> 77  077 -1.48547937 -1.567524854
#> 78  078 -0.30295372 -0.693737815
#> 79  079  0.15871970  0.045938386
#> 80  080 -1.83342942 -1.685876710
#> 81  081 -1.34528626 -0.601859973
#> 82  082 -1.67320389  0.118154231
#> 83  083  1.11620550 -1.186491780
#> 84  084  0.42791159 -0.523229045
#> 85  085 -0.47154788  0.108813107
#> 86  086  0.15596525  0.597690324
#> 87  087  0.44651231 -1.345217901
#> 88  088  0.69644115 -0.147149082
#> 89  089 -0.42254980 -0.409680727
#> 90  090  0.87991984 -0.239930445
#> 91  091  0.47904036  1.048145672
#> 92  092 -1.18422181 -0.044873159
#> 93  093 -0.17455485 -0.474778332
#> 94  094 -1.58937893  0.117027447
#> 95  095 -0.87339372 -0.381974433
#> 96  096  0.29266978 -0.156546977
#> 97  097  0.23609401  0.181235444
#> 98  098 -0.36874136  0.427877789
#> 99  099 -2.32101421 -1.233675457
#> 100 100  0.19147044 -0.622140856