Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4941284
#> Y_2 0.4941284 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.80038750  1.4397960
#> 2 -0.76839129  0.3339212
#> 3  0.06480239 -0.7745818
#> 4 -1.37373930 -0.5114284
#> 5 -0.96685114 -0.4205283

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.43344983  0.58743626
#> 2   002  0.82019189  0.13516119
#> 3   003 -0.17985497  0.52443678
#> 4   004  2.11054570  2.08362135
#> 5   005  0.62061144 -0.09793046
#> 6   006 -0.23751715 -1.27840529
#> 7   007 -0.03811802  0.17333707
#> 8   008 -1.06955418  1.57861258
#> 9   009  2.37188899  2.50222587
#> 10  010 -0.26449624 -0.23654311
#> 11  011  0.71220945  1.02815257
#> 12  012 -1.47333852  0.11162377
#> 13  013  0.30434625  0.39320231
#> 14  014 -0.54701984 -1.45713108
#> 15  015 -0.40533066 -0.96511746
#> 16  016 -2.16477186 -2.49261034
#> 17  017  1.86883903  1.10501740
#> 18  018  0.81311884  0.68285804
#> 19  019 -0.33859908  0.30152442
#> 20  020  1.60569075  1.50372701
#> 21  021  1.16511782  0.66452228
#> 22  022 -0.89970177 -1.74234353
#> 23  023 -0.43496479  0.29414727
#> 24  024  0.52832179 -1.22577069
#> 25  025 -0.85630507 -0.37897220
#> 26  026 -0.12263083 -0.43839876
#> 27  027  0.27414779  0.34003253
#> 28  028  0.28484595  0.69267417
#> 29  029  1.07791256 -0.13519876
#> 30  030  0.45318980 -0.95198943
#> 31  031  0.67347288  0.88009884
#> 32  032 -0.67325298 -0.03966577
#> 33  033  0.91543007  2.46309399
#> 34  034 -0.78240214  0.02976108
#> 35  035  0.48977106  1.48902345
#> 36  036  0.92925973  0.04106510
#> 37  037 -0.06297090 -1.00792988
#> 38  038 -0.20266914 -0.08252016
#> 39  039  0.55119479  1.35273318
#> 40  040 -0.25534492  0.36340751
#> 41  041 -1.83301166 -0.92727651
#> 42  042 -0.94507071 -0.51907039
#> 43  043 -0.28470042 -0.66590813
#> 44  044 -1.42900702 -1.60796659
#> 45  045  0.07085590  1.92996405
#> 46  046  0.85279945 -0.27514584
#> 47  047 -0.67318464  0.20700706
#> 48  048 -0.09109499 -0.36841925
#> 49  049  0.56104878  0.43525602
#> 50  050  2.01692746  0.61423031
#> 51  051  1.78551065  1.79530100
#> 52  052 -0.44057290  0.65925207
#> 53  053 -2.41979829 -0.62100270
#> 54  054  0.50026406  0.54445989
#> 55  055  0.86168662 -1.37043093
#> 56  056 -0.21974289  0.77481397
#> 57  057 -0.17828874  0.34244695
#> 58  058 -0.13475629  1.81064304
#> 59  059 -1.26166708 -0.39147329
#> 60  060 -0.43419089 -1.53278669
#> 61  061 -0.20114665  0.78653637
#> 62  062 -0.61843857 -1.27802079
#> 63  063 -0.54636442 -0.15865718
#> 64  064  1.10396951  2.75738925
#> 65  065  0.22325938  1.83450910
#> 66  066  1.42549891  1.86293635
#> 67  067  0.44154062  1.05041543
#> 68  068  0.84795425  0.61807421
#> 69  069 -1.59697693 -1.74990786
#> 70  070  1.93571465  1.05844366
#> 71  071 -1.64379967 -0.75314816
#> 72  072  0.58032016  1.76235049
#> 73  073 -1.29372335 -2.01571834
#> 74  074 -0.59453773 -0.93445877
#> 75  075 -1.39219306 -1.78946435
#> 76  076  0.79674836 -0.06730308
#> 77  077 -0.27581632 -0.55364765
#> 78  078 -1.46482764 -1.13125979
#> 79  079  1.94416144  0.54921351
#> 80  080 -0.25839800  0.81699159
#> 81  081  2.05672314  1.09996247
#> 82  082  0.57237088 -0.81745997
#> 83  083  0.43163164  0.26221065
#> 84  084  1.04572186  1.20687327
#> 85  085  0.20881194  0.51409509
#> 86  086  0.17892477  1.02748969
#> 87  087  0.39999050 -0.14649804
#> 88  088 -0.07325825 -0.32435580
#> 89  089  0.19367044 -1.58012319
#> 90  090 -1.51280476 -2.42466513
#> 91  091  0.75153496  0.93132043
#> 92  092 -1.35094113 -1.48416500
#> 93  093 -1.07257023 -2.20595571
#> 94  094  0.03172965  1.84275400
#> 95  095 -0.69440719  0.85433856
#> 96  096 -0.80156704  0.05275651
#> 97  097  0.33391201  1.00565818
#> 98  098  1.80784785  0.23154591
#> 99  099 -1.07187598  0.49578071
#> 100 100 -0.21632518 -0.78218759

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.094626224 -0.44706813
#> 2   002 -1.034302722 -0.22890237
#> 3   003  1.587873103  1.83795434
#> 4   004  1.396282331  1.27679272
#> 5   005  1.373864990  0.89516520
#> 6   006 -0.285432830 -0.28061364
#> 7   007 -0.754316635  0.34545866
#> 8   008 -0.411925736 -0.31425448
#> 9   009 -0.249629314  1.04352744
#> 10  010  1.569176467  0.39606550
#> 11  011 -0.864133457 -1.40592667
#> 12  012  0.769551107  0.70690082
#> 13  013  1.154000620  2.73958151
#> 14  014  1.236078992  0.78711429
#> 15  015 -0.024615988 -0.11621743
#> 16  016 -1.860759148 -1.74320828
#> 17  017  0.396701337  0.51927064
#> 18  018  0.032494735 -0.31858154
#> 19  019 -2.489773810 -1.83745980
#> 20  020 -1.182887221 -0.19422802
#> 21  021  0.358245063  0.56074959
#> 22  022 -1.675502460 -1.42724472
#> 23  023  0.128725866 -0.25403225
#> 24  024 -0.151972323  0.10168630
#> 25  025  1.922579716  0.97752776
#> 26  026 -0.382774783 -1.63291542
#> 27  027  0.572095168  0.22500543
#> 28  028 -0.448780495 -0.80561779
#> 29  029 -0.775182226  0.95772446
#> 30  030 -1.137910504 -1.52848784
#> 31  031 -0.189983228 -0.01101093
#> 32  032 -0.617770455 -0.24833764
#> 33  033 -0.695165922  0.82361548
#> 34  034 -0.308460195 -0.78021440
#> 35  035 -1.045603379 -0.88871597
#> 36  036 -0.987531051  0.42681702
#> 37  037 -0.222082019  0.93054419
#> 38  038  2.196373992  1.97445940
#> 39  039 -2.114916186 -1.83572087
#> 40  040 -1.582051512 -1.07461602
#> 41  041  0.221339919 -0.42543655
#> 42  042  0.242450475  0.45156610
#> 43  043  0.405390753  0.06463014
#> 44  044  0.942851054  0.91744141
#> 45  045  0.677202202 -0.26777447
#> 46  046  1.187826977  0.83846254
#> 47  047 -0.220617679 -0.21253876
#> 48  048 -1.490648809 -1.22249621
#> 49  049  1.072236473  1.46269106
#> 50  050  1.193763894  1.99760113
#> 51  051  0.265559719  1.25040260
#> 52  052  0.653365451  0.61377382
#> 53  053 -0.712164819 -0.17300648
#> 54  054 -0.557896242 -0.11885992
#> 55  055 -1.706745921 -1.03407709
#> 56  056  0.523997093  0.32810692
#> 57  057 -1.237996292 -0.98995682
#> 58  058  0.318936524 -0.04997317
#> 59  059 -1.327580915 -0.45319969
#> 60  060  1.201856995  2.41857395
#> 61  061  1.832425605  1.64472378
#> 62  062 -1.348372077 -0.18261466
#> 63  063 -1.088735961 -0.94166576
#> 64  064 -2.001051336 -1.65425574
#> 65  065 -0.950832022 -0.96650235
#> 66  066  0.697352342  1.89025605
#> 67  067 -0.288307995  0.57721925
#> 68  068  2.030439622  1.00163646
#> 69  069  0.712905215 -0.80045575
#> 70  070  0.257984498  2.04531378
#> 71  071 -0.698304367 -0.22223426
#> 72  072  1.488411234  2.38731121
#> 73  073 -1.489963837 -0.03891736
#> 74  074 -0.023438272 -0.16884110
#> 75  075 -1.184034363 -1.65990460
#> 76  076  0.284164835 -0.07106556
#> 77  077 -0.982759480 -2.04250541
#> 78  078  1.014830742 -0.29416671
#> 79  079  0.649740673 -0.58288204
#> 80  080 -0.463986306 -1.63207300
#> 81  081 -0.107063718 -0.25239191
#> 82  082 -0.726842355  0.33533247
#> 83  083  2.232101191  1.54327465
#> 84  084 -0.073336469  0.82192507
#> 85  085 -0.565632480 -0.27136707
#> 86  086  1.253725351  1.92194830
#> 87  087 -1.001197331  0.15732237
#> 88  088  0.507694352  1.12205767
#> 89  089 -1.411219470  1.12114694
#> 90  090  0.007776965  0.09601542
#> 91  091  0.389882102  0.04129760
#> 92  092 -0.585510943  1.40509782
#> 93  093 -0.331711742  0.47183018
#> 94  094  0.584443968  0.26051997
#> 95  095  0.891818922  2.03936620
#> 96  096  1.257617559  2.80393241
#> 97  097 -0.046540723  0.48957047
#> 98  098 -0.078785412  0.25361971
#> 99  099 -1.219779740 -2.73047359
#> 100 100  0.344093629 -0.21246098