Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5290591
#> Y_2 0.5290591 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.5730837  2.2210751
#> 2 -0.9524009 -0.5519218
#> 3  0.1210528 -0.5041612
#> 4  0.7481530 -0.9064605
#> 5 -0.1304318 -0.6143079

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.01201090  0.384260781
#> 2   002 -0.02802943 -0.323942042
#> 3   003  0.69372477  0.476880926
#> 4   004 -0.26432253  1.219560376
#> 5   005 -0.30370398 -0.481236678
#> 6   006 -0.35581311 -0.092027942
#> 7   007  1.39979376  1.599832252
#> 8   008 -1.44757926 -0.196051123
#> 9   009  1.26327988  0.996342827
#> 10  010 -0.55972144 -1.310367676
#> 11  011  0.25780889 -1.144475824
#> 12  012  0.38953393 -0.405878619
#> 13  013 -0.76303669 -0.304965329
#> 14  014  1.56691666  1.230403528
#> 15  015  1.20603053  1.067266528
#> 16  016  0.01287754 -0.952519379
#> 17  017  0.82974533  1.951312575
#> 18  018 -0.90040360  0.886607432
#> 19  019  0.03597900  0.686531796
#> 20  020  0.56033891 -0.007259576
#> 21  021  1.04667149  0.896919064
#> 22  022 -0.42888072  0.622926187
#> 23  023 -0.03392871  1.375318312
#> 24  024  1.14019396  1.886541616
#> 25  025 -2.27098451 -0.874482468
#> 26  026  0.03452750  0.332822470
#> 27  027  1.42563215  0.056678526
#> 28  028  0.94944753  2.638117341
#> 29  029 -0.10428532  0.255643871
#> 30  030  0.33715070 -0.549422828
#> 31  031 -0.04627467 -0.514352537
#> 32  032 -0.30435645 -1.002576740
#> 33  033 -0.34171874 -0.104706230
#> 34  034 -0.97103951  0.073668002
#> 35  035  1.61882145  1.734687751
#> 36  036  0.22890319 -0.457587867
#> 37  037  1.02622592  1.329682101
#> 38  038 -0.27748861 -0.097744482
#> 39  039  1.17795563  1.354182509
#> 40  040 -0.70922245  0.233793232
#> 41  041 -0.75058654  0.585355426
#> 42  042  0.79819182  0.262247371
#> 43  043  2.61219063  1.321373665
#> 44  044  1.12615244  1.486972955
#> 45  045 -0.55546122  0.102846499
#> 46  046  1.00987225  1.140768421
#> 47  047  0.18440063  0.998870126
#> 48  048  1.43702415  1.814516365
#> 49  049 -1.89680949 -1.611530150
#> 50  050  0.34650459  1.341537308
#> 51  051 -0.97022559 -1.808954357
#> 52  052  1.57908949  1.052122561
#> 53  053  0.10212068 -0.001000328
#> 54  054  0.55010001 -0.793577395
#> 55  055 -0.83755130  1.303957895
#> 56  056  0.59317197  0.257323039
#> 57  057  0.15836651  0.512251353
#> 58  058  0.36294320  0.041518300
#> 59  059  0.19815573 -0.539400685
#> 60  060 -0.12267272  0.977652781
#> 61  061  0.78274950 -1.354711737
#> 62  062 -0.29997439  0.158307675
#> 63  063  0.20030496 -1.101781367
#> 64  064  0.61361169  0.044510976
#> 65  065 -2.03677569 -0.573964438
#> 66  066  0.60695136  1.088341031
#> 67  067  1.31284129  0.357241001
#> 68  068  0.26891809 -0.961339971
#> 69  069  0.45480259  1.158202191
#> 70  070  1.31526343  1.123937704
#> 71  071  1.00673573  2.063255378
#> 72  072 -1.11511777  0.381705973
#> 73  073  1.11581728 -0.114808081
#> 74  074 -1.37197180 -0.861757764
#> 75  075 -0.64179867 -1.332573788
#> 76  076 -0.84255844 -0.496669395
#> 77  077 -0.22516068 -0.217388592
#> 78  078  0.13357745 -0.454406017
#> 79  079 -0.54988962  1.025803397
#> 80  080  0.13591559  0.483646722
#> 81  081  1.29046909 -0.020681240
#> 82  082  0.50929686  0.820576095
#> 83  083  0.23586116  0.955523838
#> 84  084 -0.45692031  1.326480152
#> 85  085  0.50750614 -0.545988556
#> 86  086  0.59828013 -1.478178517
#> 87  087  0.65045282  1.066486059
#> 88  088  0.39187996  0.574364378
#> 89  089  1.58392519  0.114860364
#> 90  090 -0.57621450  0.416024062
#> 91  091 -1.18910343 -1.283129147
#> 92  092  0.26123117  1.377985484
#> 93  093  0.30742991  1.290137024
#> 94  094  0.42742401 -0.028478553
#> 95  095  0.34813331 -0.232807161
#> 96  096 -1.86057608 -0.858950554
#> 97  097 -0.27665482 -1.525198967
#> 98  098 -1.81452374 -0.135246290
#> 99  099  2.74382167  1.966690401
#> 100 100  1.19021369  0.270340881

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001  0.28740120 -0.5607710123
#> 2   002  0.50307401  1.0475819414
#> 3   003  2.22619020  1.1856256170
#> 4   004 -0.07084787  0.3058468748
#> 5   005  1.57911494  0.8940229958
#> 6   006 -1.06050135 -0.2512436988
#> 7   007  1.09539968  0.0819375402
#> 8   008 -0.13500030 -1.3107903435
#> 9   009  0.14731725  0.2195693411
#> 10  010 -0.58631374 -0.6252088720
#> 11  011 -0.16248490 -0.8969639401
#> 12  012 -0.93922138  0.8796975526
#> 13  013 -0.79950021 -0.1415818073
#> 14  014 -1.32876596 -1.4098544415
#> 15  015  0.85444173  0.4867825070
#> 16  016 -0.74644731 -2.0407356881
#> 17  017  1.80823317  1.0329033182
#> 18  018  1.22994937  0.4804680515
#> 19  019 -1.39806328 -1.8805747771
#> 20  020 -1.43401565 -0.3531895883
#> 21  021  0.71335392  1.1721035652
#> 22  022  0.58012506  0.2564174994
#> 23  023  2.73226973  1.0437525665
#> 24  024  0.55299716  1.1803498789
#> 25  025 -0.92274302  0.2635255045
#> 26  026  1.16615197  1.0345080570
#> 27  027  1.20752927  1.6015013555
#> 28  028  1.29780420  0.1501798056
#> 29  029  0.38946121 -0.2723984692
#> 30  030  0.37264124 -0.0939013361
#> 31  031  0.57917421  1.7044758409
#> 32  032  1.15538416 -0.0509434428
#> 33  033  0.73976196 -1.0894492220
#> 34  034  0.89804703 -0.2110108657
#> 35  035 -0.13106497  0.2182303817
#> 36  036  1.31393888  1.2640532916
#> 37  037 -0.04435856  0.8275239436
#> 38  038  0.80741929  0.4603735139
#> 39  039 -0.47414186 -1.2461120419
#> 40  040  0.81958045  1.3234444023
#> 41  041  1.08016864  0.6610126867
#> 42  042  0.86648734  0.0656354217
#> 43  043  1.37707409  1.4764321213
#> 44  044 -0.75782657  0.2653233150
#> 45  045 -0.37416460  0.2840129433
#> 46  046  2.48865885  1.1509980379
#> 47  047  0.96785436 -0.4307316347
#> 48  048  0.99080612  0.0751111760
#> 49  049 -1.26626994 -0.1303134683
#> 50  050  0.95584167 -1.2872988540
#> 51  051 -1.35308968 -1.6848722809
#> 52  052  0.14003638 -1.2888648810
#> 53  053  1.47601473  0.2168912995
#> 54  054 -0.56778403  0.5613928293
#> 55  055 -0.51811304 -0.4306438485
#> 56  056  0.08277064 -0.0391347874
#> 57  057  0.63003942  0.4162598566
#> 58  058  0.31756901  0.3331994049
#> 59  059 -1.49989547 -1.0896894911
#> 60  060  1.95708043  0.7033613119
#> 61  061  0.26940080 -0.1831352782
#> 62  062  0.39832053  0.5009777086
#> 63  063  1.64164702 -0.9737660482
#> 64  064 -2.25675430 -2.1404982203
#> 65  065  0.12706108 -0.3646119890
#> 66  066 -2.15143672 -1.2791588357
#> 67  067  0.85008831  1.3499605910
#> 68  068  0.81406525 -0.6344477089
#> 69  069 -0.59953624 -0.2053485118
#> 70  070  0.99627015  0.0004519396
#> 71  071 -1.86901884 -0.8838749838
#> 72  072  0.47239614  1.6610032333
#> 73  073  1.05501341 -0.2494067564
#> 74  074  0.47063509 -1.0232442727
#> 75  075  0.57236717  0.4036271108
#> 76  076 -1.62795941  0.2251288187
#> 77  077 -0.24460523  0.3713612836
#> 78  078  1.29507332  2.5887966501
#> 79  079  0.24587291  0.4113474272
#> 80  080  0.86985568 -0.1866304620
#> 81  081  1.19537150  0.8277312726
#> 82  082 -0.30541644  0.3087056292
#> 83  083  0.28576419  1.3287863731
#> 84  084  0.63492608  0.8127123855
#> 85  085 -0.20751112  0.1863270766
#> 86  086  0.33179439  0.0858951412
#> 87  087 -1.45770967 -0.2540750061
#> 88  088 -1.78627338 -2.2058587888
#> 89  089  0.03309697  0.0727710924
#> 90  090 -0.30546042  0.2453027959
#> 91  091  0.64619663  0.4958153429
#> 92  092 -1.91928759  0.0198393863
#> 93  093 -0.69421775 -0.5352499254
#> 94  094  1.69464922  1.6659871819
#> 95  095 -0.44012653 -0.4149364343
#> 96  096 -0.14294898  0.6068927986
#> 97  097  0.54557026  1.3023479668
#> 98  098  1.39425359  1.1109300755
#> 99  099  0.21381758 -0.5469818203
#> 100 100  1.47321200 -0.0563481507