Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5367407
#> Y_2 0.5367407 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1 -0.842534887 -1.0163955
#> 2  0.254076790  0.1548352
#> 3  0.212241739  0.1141680
#> 4 -0.204110829 -1.1903477
#> 5 -0.007514546  1.1023760

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.093761134 -0.421231595
#> 2   002  1.168865544  0.068977760
#> 3   003 -0.985330322 -0.567555738
#> 4   004 -0.700152082 -1.843706617
#> 5   005  0.388921738  1.095837750
#> 6   006  0.835141745 -0.524233434
#> 7   007 -1.322913050 -0.596271707
#> 8   008 -0.003668966  1.334144503
#> 9   009 -0.706713913 -0.633253453
#> 10  010 -1.117096061 -1.008532080
#> 11  011 -0.650868264 -0.142405753
#> 12  012 -1.095481821 -1.227491606
#> 13  013 -0.564612659  0.150633544
#> 14  014 -1.025249913  0.226233060
#> 15  015 -0.718583164  1.207445597
#> 16  016 -0.810854805  0.471326493
#> 17  017 -0.182909456 -0.113469606
#> 18  018 -0.870627029 -1.559053745
#> 19  019 -0.095328107 -0.943350457
#> 20  020  0.053444069  0.620379466
#> 21  021  0.402430317  0.998395854
#> 22  022 -0.172683641  0.514175602
#> 23  023 -0.885742745  0.477141458
#> 24  024 -1.095644544 -1.067224370
#> 25  025  1.549365656  1.503819961
#> 26  026  2.033402817  1.434845388
#> 27  027 -0.081631861  0.860714919
#> 28  028  0.671475687 -0.140783447
#> 29  029 -0.975964288  0.475305940
#> 30  030 -1.188356903 -1.008583538
#> 31  031  0.474641162 -0.913316602
#> 32  032  0.569722662 -1.005147857
#> 33  033  0.871387056  0.174501475
#> 34  034 -0.203409140  0.014718230
#> 35  035  0.989534796  0.123554762
#> 36  036 -1.129225942  0.061856765
#> 37  037 -0.068960051  0.005693507
#> 38  038  0.806841233  0.929285677
#> 39  039  1.084128843  1.152243802
#> 40  040  0.890143791  0.765027589
#> 41  041  0.635815727  0.728276061
#> 42  042 -0.387195902  0.704432095
#> 43  043  0.204558821 -0.515976850
#> 44  044  0.597334594 -0.849340681
#> 45  045  0.101221890 -0.855542170
#> 46  046 -0.464116370 -1.263330013
#> 47  047 -1.467322840 -0.751630714
#> 48  048  0.382141598 -0.294642246
#> 49  049 -0.173096363 -1.312577909
#> 50  050  1.084502911  0.139013829
#> 51  051 -2.198138293  0.511676687
#> 52  052 -0.355468889 -0.375600137
#> 53  053 -0.199473102 -0.375617427
#> 54  054  1.646033245  2.058922186
#> 55  055 -1.108641794 -1.164078439
#> 56  056  2.232935599  1.316132742
#> 57  057 -0.327612814  0.783566452
#> 58  058 -1.004833589  1.090190952
#> 59  059  0.031776541 -1.346875031
#> 60  060  0.802089052 -0.260770320
#> 61  061  1.266174648  1.533663771
#> 62  062 -1.933553150 -2.288699137
#> 63  063 -1.550655847 -0.819147381
#> 64  064  2.140706104  1.301452376
#> 65  065 -0.709662611  0.197711928
#> 66  066  1.023626230 -0.815744644
#> 67  067 -0.942106481 -1.583242419
#> 68  068  0.786432917  0.896922737
#> 69  069  0.139821219 -0.134610553
#> 70  070 -0.411781752 -0.175089652
#> 71  071  0.129700631  0.079362502
#> 72  072  0.098345401  1.441927852
#> 73  073 -0.859596639  0.399224580
#> 74  074 -0.013449861 -0.500555663
#> 75  075  1.093328346  1.854860801
#> 76  076  0.207062270 -0.718910052
#> 77  077  2.103780050  1.849336550
#> 78  078  0.325575034  0.769408706
#> 79  079 -0.384132501 -0.409731817
#> 80  080 -0.856762185  0.797411422
#> 81  081 -0.617887737 -0.904815073
#> 82  082 -0.312678008 -0.225912272
#> 83  083  0.373822441  0.208674955
#> 84  084  0.035862259  0.227932871
#> 85  085 -0.593595695 -0.208737637
#> 86  086  0.372699564 -0.083140353
#> 87  087 -1.183298501  0.758615888
#> 88  088 -0.460459744 -1.598246744
#> 89  089 -0.242539940  1.897326974
#> 90  090  0.444575694  0.584250113
#> 91  091 -0.160516697 -0.639209053
#> 92  092 -0.524975237  0.074681921
#> 93  093 -0.420470490  1.546151987
#> 94  094 -0.928063985 -0.455870240
#> 95  095  0.387438984  1.566616493
#> 96  096 -0.444356763 -0.522832171
#> 97  097 -0.199333468  0.326952221
#> 98  098 -0.243666866  0.365057486
#> 99  099 -1.454970940 -1.169664055
#> 100 100 -0.472911662 -0.185235500

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.73053628 -0.325099264
#> 2   002 -0.60522506  0.540904495
#> 3   003 -0.55824542 -0.191479144
#> 4   004  0.67245073  1.692705805
#> 5   005  1.05467551  1.983820385
#> 6   006 -0.45079627 -1.331859288
#> 7   007  0.88475260  0.447785961
#> 8   008 -0.77745004 -0.815605510
#> 9   009  1.82938229  1.223317376
#> 10  010 -0.04679883 -1.840306639
#> 11  011  0.52482267 -0.614766457
#> 12  012  0.53142253  2.810780553
#> 13  013  1.44119612  0.432781680
#> 14  014 -0.06856247 -0.121020742
#> 15  015  2.71620780  0.002371162
#> 16  016  0.12465434 -0.536720355
#> 17  017 -0.43704241 -0.756785826
#> 18  018  1.19943791  0.607507943
#> 19  019 -0.09468675  0.090313097
#> 20  020  0.16592787  0.412814574
#> 21  021 -1.05343732 -1.285967358
#> 22  022 -1.08526819 -0.232479164
#> 23  023 -0.09782156  0.026050483
#> 24  024  0.29916937 -1.353888641
#> 25  025 -0.56376387  1.098046855
#> 26  026  0.38419115  0.703799551
#> 27  027  0.71863362  0.476840290
#> 28  028  1.11157745  0.874558500
#> 29  029  0.53268404  0.068219130
#> 30  030 -1.09067130 -1.881971344
#> 31  031  1.39708222  0.764831791
#> 32  032  0.31338861  0.960635432
#> 33  033 -1.77597743  0.478776194
#> 34  034  1.12021995  0.545242220
#> 35  035  0.87583004  0.158964025
#> 36  036 -0.32073787 -0.701948804
#> 37  037 -0.57038667 -0.513626357
#> 38  038 -0.23040747 -0.935630670
#> 39  039 -0.61297995  0.163667661
#> 40  040 -1.14118559 -0.196684223
#> 41  041 -1.08439165 -0.286345553
#> 42  042 -0.81815995 -1.042618554
#> 43  043  0.61847093  1.157392394
#> 44  044 -0.08365051 -2.322927396
#> 45  045  0.43247636 -0.110952132
#> 46  046  0.83715801  0.899031486
#> 47  047  0.63323931  1.199943493
#> 48  048  0.11758286 -0.295933199
#> 49  049  0.25638085  0.494665588
#> 50  050 -0.05848672 -0.733792530
#> 51  051  0.61105808  0.351035911
#> 52  052 -0.60191049 -0.485537172
#> 53  053  0.34121183  0.778710780
#> 54  054  2.03358378  2.370934364
#> 55  055  1.86234969  0.897076981
#> 56  056 -1.00312930 -0.170052320
#> 57  057 -0.26799368 -1.061269550
#> 58  058 -1.23214837 -0.666083771
#> 59  059  0.01544082 -1.690988610
#> 60  060  0.15649164 -0.265851793
#> 61  061  3.35471054  0.660887698
#> 62  062  0.70740296  1.106924215
#> 63  063  0.53267396  0.016829391
#> 64  064 -0.41990083  0.890280915
#> 65  065 -0.61374371 -1.472566007
#> 66  066 -0.46185156  0.112935699
#> 67  067  1.40847272  1.714732842
#> 68  068 -0.31069164  0.292325237
#> 69  069  0.26514745 -0.603358364
#> 70  070 -0.13472936  1.372063733
#> 71  071 -1.35584188 -1.359330033
#> 72  072  0.67867151  0.296688075
#> 73  073 -0.61501495 -0.450494381
#> 74  074  0.28565507 -1.085242161
#> 75  075  1.34507416 -1.409923375
#> 76  076 -0.53207563 -0.207536527
#> 77  077 -0.35721551 -0.089070955
#> 78  078  0.39370373  0.759660968
#> 79  079  0.15923305  0.422763372
#> 80  080 -1.12428179  0.352245857
#> 81  081  2.40194104  1.618318377
#> 82  082 -0.73378372 -0.614696601
#> 83  083 -0.18124564  1.451946315
#> 84  084  0.74133658  1.217685690
#> 85  085 -1.45492811  0.583874525
#> 86  086 -0.62854875 -0.188280598
#> 87  087  1.70359942  1.061092493
#> 88  088  0.56150608 -0.862159373
#> 89  089  1.58125500 -0.978654858
#> 90  090  0.57472551  1.334026786
#> 91  091 -0.14648143 -1.840192768
#> 92  092 -0.34323485 -0.249187364
#> 93  093 -0.07795123 -0.704279871
#> 94  094 -0.37126330  0.170859867
#> 95  095 -0.80263312  0.221388904
#> 96  096 -1.40083964 -1.546814238
#> 97  097 -1.48428479 -1.112425027
#> 98  098  0.82329340 -1.571947028
#> 99  099  1.24252023  1.304251333
#> 100 100  0.94519746  0.898233542