Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5171209
#> Y_2 0.5171209 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.38896199 -0.4443726
#> 2  0.03242276  1.9817956
#> 3 -0.21564417  0.9346575
#> 4  0.88448509  1.0290788
#> 5 -1.06705241 -1.0099812

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.506762060  0.564126756
#> 2   002  0.653677863  0.640703682
#> 3   003 -0.349837346 -0.359523950
#> 4   004 -2.241272190 -2.134742238
#> 5   005 -0.329333896  0.280151835
#> 6   006  0.184428107  2.500378048
#> 7   007  1.007843374  1.711535860
#> 8   008 -1.432483401 -0.364241736
#> 9   009  0.547765633  0.224360418
#> 10  010  0.298309829 -1.581347145
#> 11  011 -1.587675944 -2.119062591
#> 12  012  2.119993629  1.827075894
#> 13  013  1.727401635  1.320081262
#> 14  014 -0.784369990  0.332944047
#> 15  015  0.189668636 -1.329913503
#> 16  016 -1.275754520  0.543895737
#> 17  017  1.323593544  1.410892159
#> 18  018  1.696809199  0.896616115
#> 19  019  0.244211374 -0.581384502
#> 20  020 -0.948860891 -0.598939652
#> 21  021 -0.506687162 -0.932498362
#> 22  022  1.211013595 -0.133200364
#> 23  023  0.663996725 -0.249139975
#> 24  024  0.371153940  0.947092833
#> 25  025  0.988635984  1.029289895
#> 26  026 -0.288109058 -0.947576715
#> 27  027 -0.348839464 -0.383402826
#> 28  028 -0.223034030 -0.424414309
#> 29  029 -1.335874346 -0.371995125
#> 30  030 -0.624477740  0.633867251
#> 31  031 -0.544247347 -0.783377923
#> 32  032 -0.291108187 -0.297181548
#> 33  033 -2.355919335 -1.498891474
#> 34  034 -0.032241525 -1.842165819
#> 35  035  1.282124588  0.960580394
#> 36  036  0.197695399  0.012039858
#> 37  037 -0.225885230 -1.192172725
#> 38  038  0.804845900  1.263444293
#> 39  039  1.408658884  1.320639849
#> 40  040  1.338395692  1.030659765
#> 41  041 -0.002367493 -0.887046399
#> 42  042  1.132768048 -0.567451463
#> 43  043  0.268975477 -0.475373161
#> 44  044  1.589369635  2.097632169
#> 45  045 -0.001952368  0.248948020
#> 46  046 -1.677923977 -0.123303019
#> 47  047  0.373566531 -0.632486805
#> 48  048  0.207009145  0.577506213
#> 49  049 -0.443914965  0.936939451
#> 50  050  0.167184400  0.879708097
#> 51  051  1.176293191  0.503017408
#> 52  052 -0.846008288 -0.503202973
#> 53  053  0.441018257  0.006774189
#> 54  054  0.382833614  0.739342549
#> 55  055 -0.458010866 -0.067361414
#> 56  056  0.760562653  0.718724645
#> 57  057 -1.102108523 -1.560586774
#> 58  058 -0.640095782 -1.033288164
#> 59  059 -0.190513711  0.904371704
#> 60  060 -0.298955978 -1.608938745
#> 61  061 -0.335888682  0.792641041
#> 62  062 -0.371403747  1.041554944
#> 63  063 -0.069918068  0.615807817
#> 64  064  0.588367909  0.427758535
#> 65  065  1.239404256  0.023712647
#> 66  066  0.303245489  0.033812265
#> 67  067  0.609912660  0.695554162
#> 68  068 -0.408962114  1.592194934
#> 69  069  0.009383936 -0.847673680
#> 70  070 -0.897527447 -0.136546487
#> 71  071 -1.737083551 -0.248395265
#> 72  072  1.136377371 -0.205993769
#> 73  073  0.225401458  0.052128335
#> 74  074 -0.864377418 -1.050531665
#> 75  075 -0.488698256 -0.482460235
#> 76  076  1.608116670  0.176233454
#> 77  077 -0.658502773 -1.060054661
#> 78  078  1.285038870  0.741519629
#> 79  079  1.178825862 -0.613811553
#> 80  080 -0.713523369 -1.652197690
#> 81  081  1.579696886 -0.096608937
#> 82  082 -0.748105290  0.440352564
#> 83  083  0.701192614  1.240466809
#> 84  084 -0.845607947  0.985539888
#> 85  085 -0.759712924 -1.545722324
#> 86  086  1.137613092  1.684461523
#> 87  087  0.160271130 -0.172023459
#> 88  088  0.252722215  0.388756117
#> 89  089 -0.149081788 -0.493041342
#> 90  090  0.643172800  1.267269861
#> 91  091  0.864706551  1.544402654
#> 92  092  0.176596355  0.558514998
#> 93  093  1.065451961  1.061594215
#> 94  094  0.018258969  0.729521148
#> 95  095  3.108014944  2.438509980
#> 96  096 -0.574867483 -1.345756478
#> 97  097 -0.906661227 -0.488287552
#> 98  098 -0.028885774 -0.607244688
#> 99  099  1.421157707 -0.719825060
#> 100 100 -0.508606590  0.827240992

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.67674670  1.5528101303
#> 2   002  0.55998913  1.2424625548
#> 3   003  2.38269578  1.7077062071
#> 4   004  1.42648856  0.7794094216
#> 5   005  0.01644403  1.8608043684
#> 6   006 -1.09829956 -0.4522785335
#> 7   007 -1.89492432 -0.9970104781
#> 8   008 -0.19703740  1.0153116995
#> 9   009 -0.99318986 -1.1844325961
#> 10  010 -0.60686755 -0.8909505237
#> 11  011 -0.88341343  0.3276854000
#> 12  012  1.42093298 -0.6524892877
#> 13  013 -0.04286486  0.1194624738
#> 14  014  0.56284890  0.4068942321
#> 15  015  0.53485471  0.0165766852
#> 16  016 -0.02325009 -0.2764243732
#> 17  017 -0.82621550 -1.1006088059
#> 18  018  0.01576357  0.8595361813
#> 19  019  0.13958640 -0.7580596521
#> 20  020 -0.72405475  0.4234612245
#> 21  021  0.39362325 -0.1931371769
#> 22  022  1.46453583  1.2685322582
#> 23  023  1.48756374  0.4977030512
#> 24  024 -0.58361033  0.2360725652
#> 25  025 -0.14716952  0.6926932068
#> 26  026  0.26346882 -0.0262562859
#> 27  027  1.05769437  0.9882914992
#> 28  028  0.41446957 -0.6733969019
#> 29  029  0.35752166  0.6084080179
#> 30  030 -0.20841195  0.8273096469
#> 31  031  1.94822288 -1.3279463666
#> 32  032 -0.07599399  0.6359101965
#> 33  033 -0.01775116 -1.0438307571
#> 34  034  0.14337520 -0.3384511800
#> 35  035  1.14207601  0.5813879590
#> 36  036  0.41287329 -0.6981658739
#> 37  037  1.82406686  0.7214073679
#> 38  038  1.17737592  2.5087486188
#> 39  039  0.45949349  0.6212633758
#> 40  040  1.30638982  1.8085410549
#> 41  041  0.35215969 -0.3731193161
#> 42  042 -2.19318525 -0.4564353076
#> 43  043 -0.48804953 -0.7184087978
#> 44  044 -0.81482852 -1.5207305753
#> 45  045  1.90860872  0.8677658430
#> 46  046 -1.21673611 -1.3500924512
#> 47  047 -0.13575654  0.1435397577
#> 48  048 -0.97694988 -2.4425344624
#> 49  049  0.26374164  0.8773025740
#> 50  050 -1.44538228 -1.2413524879
#> 51  051  1.07785563 -0.5173704938
#> 52  052 -1.05401060 -1.0507516206
#> 53  053  1.16545847  2.1720427227
#> 54  054  0.80942002  0.2564577851
#> 55  055 -1.65405021 -0.1460157317
#> 56  056 -0.58232369 -0.3561330693
#> 57  057  1.08540329  0.8420046811
#> 58  058 -2.17868465 -0.2371765999
#> 59  059 -1.41656789  0.7081687738
#> 60  060  0.13791539 -0.0206196867
#> 61  061 -0.23248601 -0.2095720994
#> 62  062  0.76586065  0.6833170194
#> 63  063  1.53972501  2.3044657045
#> 64  064  0.46327356  0.3829012624
#> 65  065 -1.20122987  0.1749266359
#> 66  066  0.70366526  0.1493767371
#> 67  067 -0.74017839 -0.7883354518
#> 68  068  0.79133741  0.0034495947
#> 69  069  1.03079245 -0.3023904341
#> 70  070  1.95795434  2.3693915573
#> 71  071  1.15228966  0.5881362631
#> 72  072 -0.36873290 -0.3543716136
#> 73  073  0.82674937  0.6946773541
#> 74  074  0.85286371 -0.3993949951
#> 75  075 -1.05101431 -1.6049929438
#> 76  076 -0.47295511 -0.0154781766
#> 77  077  0.41111978 -0.8986251791
#> 78  078  0.24609544 -0.8321099033
#> 79  079  1.28628046 -0.7488901987
#> 80  080  1.21788628  1.4994565854
#> 81  081 -1.00125078 -1.7100368860
#> 82  082 -0.31105532  0.3802601408
#> 83  083 -0.71862736 -1.4271328167
#> 84  084 -0.44502503 -0.0004513922
#> 85  085  0.04701600 -0.4298635705
#> 86  086 -0.46409156 -1.0448660511
#> 87  087  0.37329276  0.7244048028
#> 88  088 -0.29931649  0.2819909799
#> 89  089  0.53039133  1.4197610224
#> 90  090 -0.15196337  0.2031330573
#> 91  091  0.38569939  0.4834583901
#> 92  092  0.60418605  0.2313784445
#> 93  093  0.23509659  0.4446452525
#> 94  094  1.15258397  0.2883303855
#> 95  095  1.44797491  1.0734458587
#> 96  096 -0.74160952 -0.0294943930
#> 97  097 -0.66113135 -0.3786158541
#> 98  098  0.18873383  0.5607377683
#> 99  099 -0.39732431  0.9399731867
#> 100 100 -0.97881670  0.1333627896