Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4809377
#> Y_2 0.4809377 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.4903307 -1.6971991
#> 2  0.9275605  1.0814386
#> 3 -0.8528168 -0.8106113
#> 4  0.9728216  0.7526866
#> 5  2.3511640  2.1724788

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.492514138  0.58152843
#> 2   002  0.877938615  0.04262365
#> 3   003 -1.786677766 -2.71714582
#> 4   004  2.248628365  0.77942631
#> 5   005  0.260270285  0.11029033
#> 6   006  0.264083483 -0.88554440
#> 7   007  0.007342222 -0.50758963
#> 8   008  0.754040357  1.52282679
#> 9   009  1.485750830  0.38077597
#> 10  010 -0.193932864 -1.17302464
#> 11  011 -0.663347016  0.74911105
#> 12  012 -0.198125883 -0.15348488
#> 13  013 -1.920056128 -2.07896747
#> 14  014  1.564301191  1.78163350
#> 15  015 -0.429498203 -0.84585540
#> 16  016 -1.292183351  0.41609769
#> 17  017 -0.282604287  0.77257587
#> 18  018  0.439449624 -0.15745442
#> 19  019  1.010908929  0.60451196
#> 20  020  0.770840629  0.70377332
#> 21  021  1.738622946  1.78682349
#> 22  022 -0.892184649 -0.25084525
#> 23  023 -0.169536061 -0.21166715
#> 24  024  0.528065133 -0.52784142
#> 25  025 -0.572414162  1.44875292
#> 26  026 -0.032518798  0.22210629
#> 27  027 -0.150972758  0.49379943
#> 28  028  0.774142376  0.36162452
#> 29  029  1.016346790  0.58908763
#> 30  030  0.730698547  0.37033163
#> 31  031 -0.512208222 -0.56122194
#> 32  032  0.756921214  2.55453638
#> 33  033 -0.536623617 -1.73078169
#> 34  034  1.379909093  1.75390845
#> 35  035 -1.495951956 -2.00329823
#> 36  036 -1.066585774  0.17792470
#> 37  037 -0.184866885  0.44361078
#> 38  038  0.265295969  1.84744979
#> 39  039 -1.664171620 -0.86752946
#> 40  040  0.566910308  0.33144208
#> 41  041  0.689993000  2.31893746
#> 42  042 -0.829018211 -1.64087452
#> 43  043 -0.142540359 -1.82427450
#> 44  044 -1.546774321 -0.82373301
#> 45  045 -0.605077596 -0.09369738
#> 46  046 -0.308902724 -0.82369304
#> 47  047 -0.370214533 -0.76108445
#> 48  048  0.283307780 -0.35538130
#> 49  049  0.198168977  0.63788558
#> 50  050 -1.018866228  0.06223809
#> 51  051  1.421036268  1.35478121
#> 52  052  0.394080897  0.73968043
#> 53  053  0.278026934 -0.34154971
#> 54  054 -0.484939416 -0.26446618
#> 55  055  0.758661164  0.49667416
#> 56  056  1.145530791  1.16548644
#> 57  057 -1.068801179 -0.17936193
#> 58  058 -0.373256418  0.05288371
#> 59  059  0.483447899  0.30589593
#> 60  060 -0.189719625  0.44097459
#> 61  061  0.113394150  1.05284048
#> 62  062  1.297097482 -0.76840349
#> 63  063  2.109565230  0.55571888
#> 64  064  0.727204502 -0.72878607
#> 65  065 -1.826755029 -1.78076672
#> 66  066  1.130460795  0.37663884
#> 67  067 -1.578493372  0.03550102
#> 68  068  0.468221342  0.11483813
#> 69  069  0.298590494  0.86630771
#> 70  070  1.417006704  0.33481469
#> 71  071 -0.171553220  2.13257969
#> 72  072  0.357256536  2.21248243
#> 73  073 -2.074284256 -1.51279123
#> 74  074 -0.382681323  0.84324385
#> 75  075  0.068829064 -0.43286779
#> 76  076  0.121934587  1.36203325
#> 77  077 -0.103625270  0.48386910
#> 78  078 -0.063961088  0.17105817
#> 79  079  1.409015072  1.46668808
#> 80  080  0.446711151  0.24588666
#> 81  081  0.215720158 -0.99382570
#> 82  082  0.772753966  1.49708929
#> 83  083  0.965052381  0.62448328
#> 84  084 -0.912941281  0.58737600
#> 85  085  0.786625604  0.94776767
#> 86  086 -0.726863800  0.32323659
#> 87  087 -0.378307232 -0.64358357
#> 88  088 -1.435780423  0.18147497
#> 89  089  1.661737603  2.62834409
#> 90  090 -1.227078716  1.10203411
#> 91  091  0.459773924  1.34697299
#> 92  092 -0.131614127 -1.33830301
#> 93  093 -1.151362577 -0.36801712
#> 94  094  0.913517950  1.47840175
#> 95  095 -1.438870042 -0.04600200
#> 96  096 -0.896987285 -0.12393093
#> 97  097 -1.017628404  0.20700209
#> 98  098  0.159783775 -0.22193518
#> 99  099  0.041454569  1.25682294
#> 100 100  1.721034311  1.77896945

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.20475232 -0.131313032
#> 2   002 -0.57867186 -0.652827872
#> 3   003 -1.13868363 -1.921826272
#> 4   004  0.66276166  0.999356830
#> 5   005  0.45874834  0.413172265
#> 6   006  1.90287085  1.485046159
#> 7   007 -0.74805714 -1.793013966
#> 8   008  1.64413551  1.163274117
#> 9   009 -0.38502788 -0.497639628
#> 10  010 -0.42002765 -1.597967946
#> 11  011 -1.40919202 -0.626002754
#> 12  012 -0.69228001 -0.489545691
#> 13  013  1.71248639  0.235863672
#> 14  014  0.83543689  0.099612510
#> 15  015 -0.97972103 -0.691776853
#> 16  016  2.86849583  0.387012890
#> 17  017 -1.93994344  0.056929606
#> 18  018  0.55016368 -0.374093277
#> 19  019  1.34741691  0.499502919
#> 20  020 -0.98350375 -1.353930423
#> 21  021  0.01773631 -1.609120714
#> 22  022 -0.27056165  0.815895187
#> 23  023 -0.72743000 -0.190368927
#> 24  024  0.90245548  0.006443574
#> 25  025 -1.31458005 -0.399176743
#> 26  026 -0.34792854 -0.875340572
#> 27  027 -0.50309916  0.593472556
#> 28  028 -1.48836339 -1.140986734
#> 29  029  0.45752891 -1.296463576
#> 30  030 -0.32683312 -0.575722386
#> 31  031 -0.15214893  0.474376570
#> 32  032 -1.10918732  0.127213401
#> 33  033 -0.85921860  0.917705134
#> 34  034 -0.54129355  0.216742897
#> 35  035 -0.32048374  0.689964245
#> 36  036  0.28111362 -0.040075555
#> 37  037  0.07187139  1.472164379
#> 38  038 -0.20302739 -1.070525261
#> 39  039 -0.85317930  0.841852568
#> 40  040 -1.91477692 -1.087042508
#> 41  041  1.07697557 -0.282240449
#> 42  042 -1.15520953 -0.371430461
#> 43  043  0.42348004  0.543697102
#> 44  044 -0.08008073 -0.527807656
#> 45  045  0.32909324 -0.040379862
#> 46  046 -0.48582028  0.179549407
#> 47  047 -0.67332473  1.498979360
#> 48  048  0.60449854 -0.126887659
#> 49  049  0.39788520  1.169438419
#> 50  050 -0.58363342 -0.365058515
#> 51  051 -1.05977411 -0.319326996
#> 52  052 -0.77250025 -0.882716741
#> 53  053  0.16280502 -1.446758842
#> 54  054 -0.05482393  0.514110954
#> 55  055 -0.93838729 -0.503638059
#> 56  056 -1.62647202 -0.345983980
#> 57  057 -0.31269681  1.767826090
#> 58  058 -1.51544443 -1.409973885
#> 59  059 -1.18671576 -0.957136798
#> 60  060  1.47606391  0.185541803
#> 61  061  2.03911247  0.909413478
#> 62  062  0.68867909  0.380564394
#> 63  063 -1.34450448 -1.290409562
#> 64  064 -1.06409450 -0.183822269
#> 65  065 -1.26791237 -0.278880662
#> 66  066 -1.59520250 -1.226075904
#> 67  067  0.16438273 -0.034310359
#> 68  068  2.03271361  1.297984599
#> 69  069  0.40522096  0.279254178
#> 70  070 -0.44568492 -1.607040707
#> 71  071 -0.21440898 -0.884611304
#> 72  072  0.46822110 -0.358450732
#> 73  073 -0.21472208 -0.667015319
#> 74  074 -0.25120674 -0.821553729
#> 75  075 -0.15833423 -0.071292586
#> 76  076  0.02411374 -0.665617139
#> 77  077 -0.96116331 -0.404911227
#> 78  078 -0.92501524 -1.453747550
#> 79  079  1.14685509  0.485396045
#> 80  080 -0.80864598  0.398555121
#> 81  081  1.28302853  0.912857179
#> 82  082 -0.38529005 -0.228031598
#> 83  083  0.16716437 -0.003746022
#> 84  084  1.08846843  0.953475034
#> 85  085 -0.16116370  0.104079892
#> 86  086 -2.09400413 -1.881879196
#> 87  087  0.98825176 -0.346455676
#> 88  088 -0.19793332 -0.666670438
#> 89  089  0.00887786  1.060645507
#> 90  090 -0.18580922  0.668101069
#> 91  091  0.24061942  1.325212505
#> 92  092 -0.93629527 -1.011059821
#> 93  093  2.57853175  0.390753712
#> 94  094 -0.07413772  0.213395929
#> 95  095  0.12477544 -0.134665314
#> 96  096 -0.08355488  1.501884089
#> 97  097  1.28074733  0.062819684
#> 98  098 -0.03596110  0.973569923
#> 99  099 -0.17528011 -0.900770268
#> 100 100  0.27367687  1.447177958