Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4731905
#> Y_2 0.4731905 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2216783 -0.6926966
#> 2  0.6674553 -0.8942120
#> 3 -2.0628259 -0.5757602
#> 4 -0.3300713 -0.6711783
#> 5  0.9352753  0.7092876

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.29981046 -0.52569258
#> 2   002  1.01043424 -0.49872500
#> 3   003 -0.99059069  1.15927487
#> 4   004 -0.83764470 -0.96563828
#> 5   005  0.27234333 -0.40310987
#> 6   006  0.04547368  0.16693663
#> 7   007  0.53420177 -1.96222032
#> 8   008 -0.27233185  1.63578416
#> 9   009 -0.07129333  1.90120235
#> 10  010  1.13766547  1.87082970
#> 11  011 -0.05555432 -0.78575076
#> 12  012 -1.63239658 -1.65679377
#> 13  013 -0.30154681  0.23766212
#> 14  014  1.30442801  0.30678378
#> 15  015 -0.69287418  0.06331658
#> 16  016 -1.50540757 -1.01105186
#> 17  017  0.33698667  0.92003978
#> 18  018 -0.24673502 -2.10492144
#> 19  019 -0.35311590  0.34254707
#> 20  020 -0.87832027 -0.82462689
#> 21  021 -0.61508373 -0.69838045
#> 22  022  0.66377312  1.06683435
#> 23  023 -0.10036062 -0.73528746
#> 24  024 -1.29187349 -0.16031586
#> 25  025 -0.19187062  0.34540789
#> 26  026 -0.33252894 -0.21557312
#> 27  027  1.65447688  0.92930187
#> 28  028  0.65363587  0.17217142
#> 29  029  0.49974430  0.62639932
#> 30  030  1.62825061  0.58628122
#> 31  031 -0.87125431  0.09553220
#> 32  032 -0.29977209 -0.69572625
#> 33  033  0.87274154 -0.12365588
#> 34  034 -0.25874954 -0.41414333
#> 35  035 -1.14029213 -0.78765749
#> 36  036 -0.67929267 -0.14699644
#> 37  037 -0.24586324 -0.07931115
#> 38  038 -0.45950637  0.32521411
#> 39  039 -0.75378883  0.20896886
#> 40  040 -1.84454684  0.28189354
#> 41  041 -1.63017594 -2.11955049
#> 42  042  0.79356384 -0.58308123
#> 43  043  1.35971678  0.82412874
#> 44  044 -0.92491149 -0.23879483
#> 45  045  0.17878295  0.76839125
#> 46  046  0.57778234  0.22482100
#> 47  047 -0.10875536 -1.51102760
#> 48  048 -0.90860865  1.99617914
#> 49  049 -0.17680314 -0.12455264
#> 50  050  0.24914981  1.55785702
#> 51  051  1.53138070  1.01410098
#> 52  052  1.13548563  0.63577365
#> 53  053 -0.47140213 -0.14627716
#> 54  054 -1.55173377 -2.03534054
#> 55  055  0.67220722  0.92066859
#> 56  056 -1.18482611  0.29432027
#> 57  057 -2.02789796 -1.53664268
#> 58  058  1.21602300 -0.61687276
#> 59  059 -1.89536870 -0.61492254
#> 60  060 -0.42547886 -1.08328824
#> 61  061  0.49846162  0.16303529
#> 62  062  0.23311654 -0.11941393
#> 63  063  0.38833708  0.28395733
#> 64  064  0.16926669  1.11987417
#> 65  065  0.05670578  1.01360419
#> 66  066  0.84383362  1.67351441
#> 67  067  0.13847322 -1.36506617
#> 68  068 -0.01263811 -0.49286240
#> 69  069 -0.18579680 -0.10178963
#> 70  070 -0.13417073 -0.85016665
#> 71  071  0.21996733 -0.17210049
#> 72  072  0.01928004  0.52833433
#> 73  073 -1.14358284 -0.77130715
#> 74  074  0.02925380  0.57543763
#> 75  075 -1.11825062  1.60388245
#> 76  076 -0.19468115 -1.27516189
#> 77  077  1.77669047  1.58075200
#> 78  078 -0.69719796  0.58787679
#> 79  079  1.26063066  1.03820667
#> 80  080 -0.13892728  0.38054406
#> 81  081  0.25504886 -2.95416884
#> 82  082  0.17678498  0.28085291
#> 83  083  0.99898228  0.68933959
#> 84  084 -1.46661342 -0.33570306
#> 85  085 -2.76715396 -1.26894784
#> 86  086 -0.63069843 -0.61360877
#> 87  087 -0.96634198 -2.88399908
#> 88  088  1.28935881 -0.47104468
#> 89  089 -0.29971174  0.90106358
#> 90  090  0.60540842  0.26965201
#> 91  091  0.50553030  0.35283681
#> 92  092  1.92313649  0.76725144
#> 93  093  1.77954472  0.27185824
#> 94  094  1.48875180  1.15085352
#> 95  095 -0.75128724 -1.31741697
#> 96  096 -0.90047446 -0.92474732
#> 97  097  0.36875138  0.12241837
#> 98  098  1.97165059  0.95502251
#> 99  099  0.39426388  0.90324496
#> 100 100 -0.33331021 -1.38939373

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.155268009 -1.30048267
#> 2   002 -1.147924405 -0.30767727
#> 3   003  0.445570674 -0.10429876
#> 4   004 -0.131576410 -1.12079528
#> 5   005  0.639797301  2.08372311
#> 6   006 -1.649530536 -1.00938637
#> 7   007  0.580465748  0.95049699
#> 8   008 -0.108143670  0.84292439
#> 9   009  1.033798478  0.54250080
#> 10  010 -0.343450605 -1.29529802
#> 11  011  1.241162563  0.56130942
#> 12  012 -1.190507428 -2.04112126
#> 13  013  0.118122541  0.30669901
#> 14  014  0.519908705 -0.51923014
#> 15  015 -0.078856979  0.67610496
#> 16  016  0.055212204 -0.44582438
#> 17  017  0.517457764  0.94356006
#> 18  018  1.040699188 -0.47815597
#> 19  019 -0.379747162 -0.15062893
#> 20  020  0.375984803 -0.66344436
#> 21  021  0.421612547 -0.09269465
#> 22  022 -1.886136517 -0.17226794
#> 23  023 -0.813895415 -1.28377209
#> 24  024 -1.475078458 -2.09627897
#> 25  025  0.748459069 -0.11974939
#> 26  026  1.192686613  1.42841673
#> 27  027  0.633825189 -1.19646076
#> 28  028  0.447524326  1.02343085
#> 29  029  0.430844875  0.54563838
#> 30  030 -0.002282215  1.07175211
#> 31  031  1.079737135  0.38287714
#> 32  032  0.847966069 -0.05889742
#> 33  033  1.086588606 -0.61402994
#> 34  034  0.341497011  0.93687981
#> 35  035  0.317809868  0.55453243
#> 36  036 -1.136232508 -0.13271779
#> 37  037 -1.011178852  0.92303402
#> 38  038 -0.993216807 -0.48989145
#> 39  039  0.466584207  0.69635204
#> 40  040 -0.299598360  0.63602868
#> 41  041  0.313260335  0.99060578
#> 42  042 -0.354348225  0.19691381
#> 43  043  0.782298077  0.06350712
#> 44  044  1.109759146 -1.22248455
#> 45  045  0.237467778  0.31465843
#> 46  046 -0.978011044  0.15526605
#> 47  047 -2.032548596 -1.94166912
#> 48  048  0.146691324 -1.08270064
#> 49  049  0.824397151  0.19884350
#> 50  050  0.339069447  0.34405694
#> 51  051 -1.383301234 -1.04909958
#> 52  052 -1.439116318 -0.32107385
#> 53  053  0.144901513 -0.57868479
#> 54  054  0.302065872  1.02193573
#> 55  055 -0.315101371  1.54405709
#> 56  056 -0.714981515 -1.07462720
#> 57  057 -0.843699789 -1.30422478
#> 58  058  0.401323700 -0.12138006
#> 59  059 -0.203781874  0.02620039
#> 60  060  0.846866721 -0.52364930
#> 61  061 -0.927734198 -0.14683371
#> 62  062  0.086762365 -1.15089472
#> 63  063 -0.533544601 -1.23346151
#> 64  064  0.038627310  0.37231749
#> 65  065  1.277559659  1.78145848
#> 66  066  0.159000291  1.34338685
#> 67  067 -1.039665021 -0.83354308
#> 68  068  0.379629618  0.26075215
#> 69  069  0.302658412 -1.08237030
#> 70  070  1.473648693 -1.02663251
#> 71  071  0.136392526 -0.10586693
#> 72  072  0.502490531 -0.92329828
#> 73  073 -0.371581005 -1.98535803
#> 74  074 -1.045834732 -0.90861826
#> 75  075 -1.497125158 -1.56617539
#> 76  076 -1.546723726 -0.66162345
#> 77  077 -0.676778129 -0.40189287
#> 78  078 -1.115289934 -0.41042841
#> 79  079 -1.700866456 -0.26069001
#> 80  080  1.437684875  0.64265976
#> 81  081  0.311173997  0.70050356
#> 82  082 -1.030536141  0.20247337
#> 83  083  1.429675153  1.02717998
#> 84  084  0.089269525 -0.82890286
#> 85  085  0.437571323  0.48558086
#> 86  086 -0.434049131 -2.11743354
#> 87  087 -2.422239746 -1.03419223
#> 88  088 -0.690712639 -0.55838431
#> 89  089 -0.585314715 -2.25318745
#> 90  090 -1.038054458 -0.55948160
#> 91  091 -1.006925278 -0.59528883
#> 92  092 -1.584400554  0.97983049
#> 93  093  0.670542530  0.48995719
#> 94  094  0.422857691  0.64688457
#> 95  095 -0.058130447 -1.08517291
#> 96  096  1.352347146  1.73403379
#> 97  097  0.301676601  0.41711207
#> 98  098 -0.623156686  1.35524821
#> 99  099 -0.286197250 -0.71300370
#> 100 100 -0.415286204  0.98658893