Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5272799
#> Y_2 0.5272799 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1       Y_2
#> 1 0.5397040 0.8188228
#> 2 0.1947836 0.9999016
#> 3 0.2061715 0.2506377
#> 4 0.7516401 0.4797855
#> 5 0.4201800 1.6479203

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.034306576 -0.94478755
#> 2   002 -2.194854080 -0.89191453
#> 3   003 -0.313139100 -0.82640721
#> 4   004  0.051675996  1.18802982
#> 5   005  0.178939265 -0.48229163
#> 6   006  0.705145950 -0.89654740
#> 7   007 -0.914018206 -2.26249458
#> 8   008 -1.149049231 -2.18806319
#> 9   009 -0.906763996 -0.17933522
#> 10  010  0.385945724  1.74228788
#> 11  011  0.751169097 -0.44271623
#> 12  012 -0.112008190 -0.93920886
#> 13  013 -0.467338298 -1.40990601
#> 14  014  0.368204845  0.27935319
#> 15  015  0.380339313 -0.37795887
#> 16  016 -0.151576355 -0.14097675
#> 17  017 -1.745241630 -1.10641141
#> 18  018  0.056794188 -0.14557767
#> 19  019 -0.994014415 -1.16072434
#> 20  020  0.026204955 -0.39210307
#> 21  021 -0.013790222  1.13091919
#> 22  022  0.406196416  0.19349971
#> 23  023  0.215815719  0.51877398
#> 24  024  0.336186401  0.40892812
#> 25  025 -0.547164950 -0.37409245
#> 26  026 -0.567388732 -0.54537998
#> 27  027  0.601041731  0.56846745
#> 28  028  0.014911137 -0.45407362
#> 29  029 -0.378315506 -1.27501902
#> 30  030  0.910447796  1.67035841
#> 31  031 -0.767122004 -0.82842842
#> 32  032  1.289731960  2.12805287
#> 33  033  0.622993914  0.17670800
#> 34  034  0.060073252  0.39835053
#> 35  035 -0.260974262 -0.40731429
#> 36  036  0.291856770  0.20471147
#> 37  037 -0.660623536  0.54853373
#> 38  038  0.098908096  0.30379743
#> 39  039  1.616117062  2.68475454
#> 40  040  0.753167496  0.75040871
#> 41  041 -1.085171109 -1.14963846
#> 42  042 -1.221495123 -1.53866063
#> 43  043 -0.004649961  1.85542795
#> 44  044 -0.061462104 -0.45106605
#> 45  045 -0.686486035 -0.75726265
#> 46  046  0.733677759 -0.27411247
#> 47  047 -0.157526761  0.37340938
#> 48  048  1.601674267  1.09489213
#> 49  049  1.643470374  2.54914146
#> 50  050  1.068949029 -0.77609198
#> 51  051  2.762772647  0.24758606
#> 52  052 -0.050956124  2.12167961
#> 53  053 -0.621772394 -0.40240611
#> 54  054 -0.210051550 -0.16718460
#> 55  055 -0.596032407 -1.65072646
#> 56  056  0.444677094  1.32049405
#> 57  057  0.712395774  1.37534810
#> 58  058 -0.747353615 -0.63371475
#> 59  059 -0.204094487  1.77779224
#> 60  060  2.150519772 -0.07893169
#> 61  061  0.970507895 -0.03205569
#> 62  062  0.768715028 -0.87521581
#> 63  063  1.287051433 -0.20522629
#> 64  064 -0.520122715  0.20110138
#> 65  065  0.443177422  1.26183776
#> 66  066  0.856794683  0.87226712
#> 67  067  2.232053656  2.08728892
#> 68  068  0.473116698 -0.76432677
#> 69  069 -1.285257988 -0.78751915
#> 70  070  1.182920172  0.98561806
#> 71  071 -0.068911361  0.50779463
#> 72  072 -0.967389292  0.17778364
#> 73  073  1.475373116  1.10197713
#> 74  074  0.787459514 -0.45449573
#> 75  075 -1.058577035 -1.26355974
#> 76  076  1.430567011  1.07643671
#> 77  077 -0.755728141  0.73488886
#> 78  078 -0.206221039 -0.87592404
#> 79  079 -0.901929166  0.81387738
#> 80  080 -0.263605375 -0.09021234
#> 81  081  0.101497748  0.97125795
#> 82  082  0.439679130  0.28106051
#> 83  083  0.218114590  0.69272701
#> 84  084 -0.088475784  0.50607022
#> 85  085 -0.855209369 -0.32142731
#> 86  086  0.703982165  1.09158347
#> 87  087 -1.091908772  0.71395359
#> 88  088 -2.125688678 -1.60395720
#> 89  089 -0.140117946  0.35475031
#> 90  090 -2.454067212 -1.52878841
#> 91  091  1.781481450  0.39803367
#> 92  092  0.473965165 -0.43641608
#> 93  093  0.208307041  1.19075488
#> 94  094 -0.481844474  0.64371035
#> 95  095  1.042655401  1.41975839
#> 96  096 -0.237067451  0.16890983
#> 97  097  1.242274069  0.84584024
#> 98  098 -1.360191318 -1.62224095
#> 99  099 -0.209500365  0.36126019
#> 100 100 -0.603606094 -0.68209877

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  2.32004754  1.06215201
#> 2   002  0.80159515 -0.83074356
#> 3   003  1.09122159 -0.41069580
#> 4   004  2.43535073  1.91304206
#> 5   005  0.43896878  1.02363546
#> 6   006 -0.02853835 -1.22880224
#> 7   007 -0.63486750  0.04366522
#> 8   008  1.56124279 -0.07271476
#> 9   009 -0.36073098 -0.57569365
#> 10  010  0.08753672 -0.15221966
#> 11  011 -0.93454585  0.13121564
#> 12  012 -0.62008731 -1.66189482
#> 13  013  0.03030704 -0.78002341
#> 14  014  2.14107322  2.38611997
#> 15  015 -0.43271511 -0.99466547
#> 16  016  0.97130685  0.20862329
#> 17  017  0.54422466 -0.90393342
#> 18  018  1.06817652  0.27247047
#> 19  019  0.64876580  1.27018442
#> 20  020  1.11789998 -0.94369880
#> 21  021 -2.08748008 -2.37823152
#> 22  022  0.04351975  1.01252016
#> 23  023  0.62488509  0.33271228
#> 24  024 -1.14579341 -0.57477183
#> 25  025 -0.85541046 -1.26146359
#> 26  026  0.59180283 -0.24831948
#> 27  027 -0.16045009 -1.31888705
#> 28  028  0.39493664  1.46246109
#> 29  029  1.38485078  0.91942649
#> 30  030 -0.62575328 -1.41752338
#> 31  031  0.19433556  1.57186690
#> 32  032 -2.12629132  0.12431999
#> 33  033 -0.08978600  1.01150169
#> 34  034 -0.21145540  0.33855422
#> 35  035 -0.50321756 -0.63748092
#> 36  036 -0.10437817  0.82419238
#> 37  037  0.51057169  0.70747832
#> 38  038  1.42569921  0.52362841
#> 39  039 -0.11523388 -0.85541197
#> 40  040 -0.33533239  0.89005956
#> 41  041  2.12398746  2.20198679
#> 42  042 -0.36314042 -0.78985760
#> 43  043 -1.16701713 -1.20389469
#> 44  044 -0.94217676 -0.08784990
#> 45  045 -1.06161194 -1.61231549
#> 46  046  0.95792894  2.17796084
#> 47  047 -0.09723008 -0.01169147
#> 48  048  1.07373859 -0.83830111
#> 49  049 -0.49482141 -0.11246838
#> 50  050  0.39575687  0.11284316
#> 51  051  0.17109807 -0.26423057
#> 52  052 -0.29019608  0.41697209
#> 53  053  0.48092255  0.47951083
#> 54  054 -0.59927948 -0.06445828
#> 55  055 -0.79807599 -0.20340003
#> 56  056  1.30410824  2.89895303
#> 57  057  0.44592358 -0.19135028
#> 58  058 -0.25707098 -0.05485439
#> 59  059  2.19396172  0.47442661
#> 60  060  0.14390662 -0.39311551
#> 61  061 -0.69335704  0.30817107
#> 62  062 -0.53517765  0.11143553
#> 63  063  0.08710840  0.58807006
#> 64  064 -2.27895809 -1.74619907
#> 65  065  0.01585779  0.65863552
#> 66  066  0.47649073  0.45164768
#> 67  067  1.27646808  0.20169438
#> 68  068 -0.68501460  0.22416054
#> 69  069  0.27623125 -0.54672635
#> 70  070  1.58318331  0.99781710
#> 71  071  0.89067509  0.10688689
#> 72  072 -0.57099205 -0.78223197
#> 73  073  0.62959760 -0.17471411
#> 74  074 -0.26561328 -0.54065390
#> 75  075  0.45392104  0.31744704
#> 76  076  1.15591379  0.99060634
#> 77  077  2.39011628  1.75928305
#> 78  078  0.32311636 -0.02445124
#> 79  079 -0.97819375 -0.55114192
#> 80  080 -0.51475200 -0.35603352
#> 81  081 -1.04290174 -1.20570391
#> 82  082  0.02522976 -0.16749797
#> 83  083 -1.35453689 -1.59454266
#> 84  084  0.11033098 -1.23210021
#> 85  085  1.05482517 -0.44621265
#> 86  086  1.03274173  0.23923431
#> 87  087 -0.28088766 -0.63331979
#> 88  088 -1.52885333 -0.91897635
#> 89  089  1.95147132  1.53062210
#> 90  090 -1.13838801 -0.29820954
#> 91  091  0.11691677  0.31721915
#> 92  092  1.40912448  0.49814635
#> 93  093 -0.05645847 -0.91461386
#> 94  094  0.17309894 -0.06138063
#> 95  095 -1.40208963  0.76509934
#> 96  096  0.32315093  1.15224633
#> 97  097  0.15576176  0.38661939
#> 98  098 -0.68344767 -1.90022755
#> 99  099 -0.12810778 -1.37769904
#> 100 100 -0.21112871  0.72634237