Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4949793
#> Y_2 0.4949793 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  0.92107715 -0.53705814
#> 2  0.34720190  1.04400514
#> 3 -0.56043116  0.06357335
#> 4 -0.06274935 -0.12063773
#> 5 -0.94154019 -0.94241571

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.632027178  0.033671361
#> 2   002  1.316665366  0.317144902
#> 3   003  1.337356014  0.723379705
#> 4   004 -1.071373669 -0.941063075
#> 5   005 -1.055338272 -0.496146877
#> 6   006  0.036555483 -0.172948918
#> 7   007 -0.930292227 -2.330267841
#> 8   008 -0.684431235 -1.297187755
#> 9   009 -1.058898655 -2.250654670
#> 10  010 -0.364563891 -0.492643258
#> 11  011  1.486173421  1.328751634
#> 12  012  0.772124641  0.440211745
#> 13  013 -0.530837795 -0.855121586
#> 14  014  0.581174865  0.291502884
#> 15  015  0.980118917  0.197153605
#> 16  016  0.053494722  1.202222240
#> 17  017 -0.381917849  1.081758272
#> 18  018 -1.465785967 -0.279265863
#> 19  019 -0.769193085 -0.829061443
#> 20  020 -1.205980916 -0.124674555
#> 21  021 -1.765125657 -0.813925805
#> 22  022 -1.655899110 -2.121520654
#> 23  023 -2.224603023 -2.084123133
#> 24  024  0.883635956  1.300589237
#> 25  025  0.519518549 -0.242641352
#> 26  026  0.921738838  0.830099460
#> 27  027 -1.099077536 -0.334148095
#> 28  028  1.601898632  0.331351614
#> 29  029 -1.194457523 -1.705288515
#> 30  030 -1.229655291  0.389494392
#> 31  031 -0.961997968 -0.121957140
#> 32  032 -2.458036961  0.242215766
#> 33  033  0.970566726  0.516701242
#> 34  034 -0.671490864 -0.152575589
#> 35  035  0.937889689 -0.048664864
#> 36  036 -1.308665484  0.707616855
#> 37  037  0.625090362 -0.437814584
#> 38  038  2.326258187  0.743908073
#> 39  039  0.366841784  1.265494527
#> 40  040  0.368781055 -1.057306380
#> 41  041 -0.327141020  0.130437682
#> 42  042  0.738830252  0.781050658
#> 43  043 -0.580863040 -2.140394300
#> 44  044  1.025640055  1.202203172
#> 45  045 -0.948116682 -0.124543762
#> 46  046  1.138913206  1.767287232
#> 47  047  1.166004241 -0.243568881
#> 48  048  0.336742878  1.144732089
#> 49  049 -0.198187043  0.035500293
#> 50  050  1.681344948 -0.540528189
#> 51  051 -0.444221545  1.218832181
#> 52  052  0.213073722 -0.191087847
#> 53  053  0.751214131  0.630208520
#> 54  054 -0.480091305  0.636624733
#> 55  055 -0.554452064  0.070868570
#> 56  056 -0.961129745  1.046275358
#> 57  057 -0.652544307 -1.169182167
#> 58  058  0.958086474  0.108990658
#> 59  059 -0.352979685 -0.708318511
#> 60  060 -0.250408257 -0.097207689
#> 61  061 -0.480259252  0.453432066
#> 62  062 -0.464221991 -0.546709798
#> 63  063 -0.833500737  0.138169347
#> 64  064  0.216289940  0.501196329
#> 65  065 -0.565223317 -1.125180886
#> 66  066  0.023647774  0.605725621
#> 67  067  1.064937998  0.619077965
#> 68  068  0.074524927  0.970888184
#> 69  069 -1.114382082  0.335860788
#> 70  070 -1.193931137 -2.362454480
#> 71  071  0.495166468  0.712439448
#> 72  072 -1.236855974 -0.860102519
#> 73  073  1.356311203  0.731713236
#> 74  074  0.332662561  1.038983095
#> 75  075 -0.671440792 -1.071836262
#> 76  076  0.526875004 -0.004942695
#> 77  077  0.007495303  0.237949249
#> 78  078  0.869328539  1.505153838
#> 79  079  0.922600568  0.414697436
#> 80  080  1.426830746  1.857510091
#> 81  081  0.087359809  0.813521980
#> 82  082  0.569989028 -1.249405532
#> 83  083 -1.139806859  0.029698804
#> 84  084 -1.193183487 -0.049658202
#> 85  085 -0.098921893  0.403373592
#> 86  086  0.998284462 -0.959655563
#> 87  087 -0.351253991 -1.222794545
#> 88  088  0.318225815  0.005003291
#> 89  089 -0.319322908 -0.108512176
#> 90  090  0.027461628  0.372438956
#> 91  091  1.379950371  1.114224397
#> 92  092 -1.026863925 -1.665230506
#> 93  093 -1.171858817 -3.100896728
#> 94  094  0.147335168 -0.097631973
#> 95  095  0.161188534 -0.596617615
#> 96  096  0.114151698 -0.508951761
#> 97  097 -2.096231454 -2.189716171
#> 98  098 -0.410544878 -0.540351284
#> 99  099  0.166779324  0.887926288
#> 100 100  0.501193511  0.805217522

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.49385827  0.540730089
#> 2   002  0.44447281  1.604727112
#> 3   003 -1.14478671  1.352573654
#> 4   004 -0.03273100  1.167217714
#> 5   005  1.49215123  0.571380721
#> 6   006 -0.04904074  1.072347175
#> 7   007  0.58016877  0.540616410
#> 8   008  0.13880584  1.460572951
#> 9   009 -1.00537629 -0.916943690
#> 10  010  0.59376508 -0.075367660
#> 11  011 -1.48381896 -1.031979049
#> 12  012 -0.09154641  1.527824944
#> 13  013  0.14216823  1.156743483
#> 14  014  0.98053882  0.618333489
#> 15  015  0.16123207  0.678554016
#> 16  016  0.64676066  1.505264218
#> 17  017  2.30767355  1.068340506
#> 18  018 -0.03053061  0.941447980
#> 19  019 -1.34091355 -1.666729622
#> 20  020 -1.63390970 -0.221522582
#> 21  021  1.95380154 -0.109721833
#> 22  022 -0.19854993  1.433800841
#> 23  023  0.63450066 -0.207446508
#> 24  024  0.40346927  0.253478598
#> 25  025 -0.45667727 -0.945310256
#> 26  026  0.78776942  1.423727853
#> 27  027  0.18537904 -1.259924580
#> 28  028 -0.64177271 -0.953944782
#> 29  029  1.18432599 -1.350724114
#> 30  030  0.73275004  0.587591832
#> 31  031 -0.65130502 -0.183367196
#> 32  032 -1.53038215 -1.258714605
#> 33  033  0.59634771  1.346590369
#> 34  034 -1.27718034 -3.066250877
#> 35  035 -0.30335713  1.079775801
#> 36  036 -1.47646818 -1.373455050
#> 37  037 -0.07818563  0.297748495
#> 38  038  0.61654684 -0.093744299
#> 39  039 -1.52543786 -1.939080936
#> 40  040  1.62595062  0.851912858
#> 41  041  1.13899103  0.126014321
#> 42  042 -0.41938573 -1.314354820
#> 43  043 -0.41401131 -0.278876575
#> 44  044  0.52609378  1.393541464
#> 45  045 -1.44593527  0.714904985
#> 46  046 -0.14910788 -0.035164226
#> 47  047 -0.37977561 -0.030418958
#> 48  048  0.12327438  0.082900895
#> 49  049 -0.27575423  1.142616591
#> 50  050 -0.03138400 -0.841269353
#> 51  051 -1.35655500 -2.769392249
#> 52  052 -2.09237707 -2.423108742
#> 53  053 -0.31901937  3.061888567
#> 54  054  0.79718451  1.668812316
#> 55  055  0.46537490  0.489876723
#> 56  056  0.02723225  1.202303736
#> 57  057  1.95540922  1.136422623
#> 58  058 -0.30310227 -0.205416980
#> 59  059 -0.90963736  0.801295936
#> 60  060 -1.48531638 -1.895126223
#> 61  061  1.52515177  1.261190128
#> 62  062 -0.81744247 -0.109417408
#> 63  063 -0.47450403 -0.358538343
#> 64  064  0.72364049 -2.697688321
#> 65  065 -0.07198550  0.253575117
#> 66  066  0.58286260 -0.429779269
#> 67  067  1.02025607  1.762257992
#> 68  068 -0.43424892 -1.025383162
#> 69  069 -0.99395916 -0.824871734
#> 70  070  1.62439856 -0.179850813
#> 71  071  1.46089341  2.107752151
#> 72  072 -1.52304233 -0.893089609
#> 73  073 -0.52793790 -0.827422353
#> 74  074  0.07009297  0.595610530
#> 75  075 -1.48378292 -1.238153080
#> 76  076 -0.57404746  0.307973910
#> 77  077 -1.26175055 -1.448367550
#> 78  078 -1.70210864 -0.358057423
#> 79  079  0.34668145  0.240047458
#> 80  080 -1.41857131 -1.379040171
#> 81  081 -3.03235819 -2.241737487
#> 82  082  0.22523419  0.233743538
#> 83  083 -0.62866655  1.089388301
#> 84  084 -1.06327577 -1.347564881
#> 85  085 -0.23155208  0.928420962
#> 86  086  2.23519296  1.185780516
#> 87  087 -1.06556698 -0.480593106
#> 88  088  1.23927072  0.215571903
#> 89  089 -0.70618632  0.717428852
#> 90  090  0.93878374  0.114601929
#> 91  091 -2.16970016 -0.608377292
#> 92  092 -1.71638744 -0.017795062
#> 93  093  0.18785430  1.380087697
#> 94  094  0.36102227  0.143031964
#> 95  095  0.68429431 -0.005133783
#> 96  096 -0.25394011 -0.034077293
#> 97  097  0.03237651 -0.086307806
#> 98  098  1.04703107 -0.282201798
#> 99  099  2.49603788  1.446840560
#> 100 100 -0.25061767  0.594676963