Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5438999
#> Y_2 0.5438999 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.2860920  0.4700228
#> 2  1.6729408  0.7063744
#> 3 -0.2033927 -0.8217171
#> 4  0.1098879  0.3261422
#> 5  0.2909479  1.1329117

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1           Y_2
#> 1   001 -0.589138599 -0.0941532944
#> 2   002 -0.008687481  0.4146738956
#> 3   003 -1.221077877 -1.1172141542
#> 4   004 -0.079230020  0.5680908899
#> 5   005  0.252766477  0.5885440635
#> 6   006 -0.496042712 -0.3882481169
#> 7   007 -0.290967219 -0.1364677270
#> 8   008 -0.983642234 -1.2091326498
#> 9   009  1.585565064  0.9153560936
#> 10  010 -0.874736444  1.0866290761
#> 11  011  0.690246330 -0.2638853731
#> 12  012 -0.367817345  1.6483008165
#> 13  013  1.317059161  1.4334932386
#> 14  014  0.084394958 -0.8593126097
#> 15  015  0.522727050  0.7379938348
#> 16  016  0.056429718  0.3891579760
#> 17  017  0.539867300  1.5104580723
#> 18  018  0.828161037  1.1278763448
#> 19  019  0.655989445  0.5051606302
#> 20  020  1.098008340  0.8961579546
#> 21  021 -0.533309904  1.0614452055
#> 22  022 -0.697389917 -0.1327259534
#> 23  023 -0.390454927 -0.8444318172
#> 24  024 -0.689280160 -0.4394702689
#> 25  025 -1.742597692 -2.6395684391
#> 26  026  0.612548017  0.4431521614
#> 27  027  0.148988345  0.8575993270
#> 28  028  0.583859670 -1.0682255618
#> 29  029  1.108322093  0.5237619225
#> 30  030  0.353710049  0.9561824104
#> 31  031  0.355539482  0.3497837741
#> 32  032 -0.634038326 -0.3343217723
#> 33  033  2.321896857  1.7264582745
#> 34  034 -0.707653682 -0.2346155221
#> 35  035 -1.053412219 -0.1471834718
#> 36  036 -1.170424079  0.0224589472
#> 37  037  0.320724751 -0.7405911121
#> 38  038 -1.374812044 -1.1163459912
#> 39  039 -0.952648674  0.2135003408
#> 40  040  0.532468110  1.0102797244
#> 41  041  1.017111032 -1.1992651132
#> 42  042  1.108943050  0.1991799191
#> 43  043 -1.172885636 -1.1196285227
#> 44  044  0.705575329  0.3700454985
#> 45  045  0.792046897  0.0982994875
#> 46  046 -0.478346060 -0.6151509525
#> 47  047  0.631529873  1.1589236710
#> 48  048 -1.503193650  0.3746106886
#> 49  049 -1.036732191 -1.4837792244
#> 50  050  0.655899633 -0.0375333166
#> 51  051  1.066042847  0.2902771622
#> 52  052 -0.220639652 -0.3935537883
#> 53  053 -1.452358089 -0.7546906596
#> 54  054 -1.061723659 -1.5905982786
#> 55  055  0.228302252 -0.0139041103
#> 56  056 -0.381293001  0.2322249112
#> 57  057 -0.508928141  0.9364564494
#> 58  058  1.710020443  0.5738497997
#> 59  059 -0.275604168  0.3802941464
#> 60  060 -0.700016591 -0.4439019358
#> 61  061 -0.343611333 -1.1558729669
#> 62  062 -2.277398329  0.1262050467
#> 63  063 -0.779233546  1.1122705679
#> 64  064  0.357263657  0.9413114413
#> 65  065  0.048828945  0.2990886198
#> 66  066  0.508354624 -0.2950568774
#> 67  067  1.173551169  0.0287969272
#> 68  068 -0.427576167 -1.9101585075
#> 69  069  0.712633665 -0.5031378846
#> 70  070 -0.340598959 -1.0872479349
#> 71  071  2.168271862  0.9304056162
#> 72  072 -0.441028879  0.0530151516
#> 73  073 -1.177809199 -0.7536706121
#> 74  074 -0.273302992 -0.1351071031
#> 75  075 -1.210017974 -0.6316276592
#> 76  076  0.739184629  0.2347240193
#> 77  077  0.056694023 -1.0769581422
#> 78  078 -0.340355985 -0.4541644539
#> 79  079  2.026413829  0.8335811967
#> 80  080  0.678306009 -0.1223302128
#> 81  081  0.134262013  0.1584778489
#> 82  082  0.845513626 -0.3205854796
#> 83  083 -1.959699741 -1.5897316278
#> 84  084 -0.268401414 -1.4449423871
#> 85  085 -0.192089663 -1.9455359353
#> 86  086  0.416497298  0.0009861896
#> 87  087 -0.803398884  1.6652306821
#> 88  088 -0.116861761 -1.5534724774
#> 89  089  1.182858203  0.8930873411
#> 90  090  0.799957940 -1.5887611067
#> 91  091 -0.227366210 -1.2782554442
#> 92  092 -0.078382154  0.9886712648
#> 93  093 -0.468929962 -0.2403868855
#> 94  094  1.846936542  0.9174439764
#> 95  095 -0.237992138 -0.4553079534
#> 96  096 -0.610457018 -0.0599938343
#> 97  097  0.427817112  0.9548075334
#> 98  098  0.342954714 -0.4750523944
#> 99  099  0.638720800  1.1838859063
#> 100 100  0.103774735  0.2443568569

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.672941207  0.53466634
#> 2   002 -1.431510017 -1.04608424
#> 3   003  0.487421537  0.46895569
#> 4   004 -1.453026202 -2.13367611
#> 5   005 -0.323487436 -0.03944623
#> 6   006 -2.188587745 -0.90889765
#> 7   007  0.360525529 -0.54580199
#> 8   008  0.474387815 -0.96938006
#> 9   009 -0.426069196 -0.85756710
#> 10  010  0.178371111 -1.17905742
#> 11  011 -1.295360435 -0.68371469
#> 12  012 -0.602722087 -1.98274044
#> 13  013 -1.573359827  0.03223912
#> 14  014 -1.283136273 -1.15474214
#> 15  015  0.137673971 -0.83185555
#> 16  016  0.569901126 -0.09895383
#> 17  017 -0.236817207 -0.92804353
#> 18  018 -1.305170579 -1.76733693
#> 19  019  0.003221372 -0.56730574
#> 20  020  0.206181129  1.24588977
#> 21  021 -0.088175776  1.07667275
#> 22  022  0.879342104  0.11390616
#> 23  023 -0.119676511  0.08507622
#> 24  024  0.448324530  0.70525336
#> 25  025  1.394553315  1.58459544
#> 26  026 -1.679348643 -0.75483442
#> 27  027 -1.030119798 -0.43444852
#> 28  028 -0.638619609  0.75442765
#> 29  029 -1.402613073 -2.59365266
#> 30  030 -0.495404438 -0.32069142
#> 31  031 -0.101934010 -1.51956241
#> 32  032  1.466413577  1.83842700
#> 33  033 -0.152608280 -0.17993956
#> 34  034  0.027247608 -0.47926423
#> 35  035  0.526047726  1.98340687
#> 36  036  0.582373060  0.53762730
#> 37  037  1.129757605  0.87168709
#> 38  038  2.581341513  2.87227259
#> 39  039  0.376115127  0.87743866
#> 40  040 -1.393726439 -0.30577421
#> 41  041  0.702921942  1.02495264
#> 42  042  0.128442488 -0.32537471
#> 43  043  0.427364503  0.26093266
#> 44  044  0.549268570  0.65774080
#> 45  045  0.623594913  0.08108611
#> 46  046  1.368138651  0.19455977
#> 47  047  1.267014286  0.56280271
#> 48  048  0.987983781  0.35969883
#> 49  049 -0.056869533 -1.35136961
#> 50  050  0.513187803  0.62379156
#> 51  051  1.344841317  0.89539433
#> 52  052  0.810746960 -0.73012767
#> 53  053 -0.315613256  1.08790260
#> 54  054  1.265158583 -1.00167286
#> 55  055  0.186288747  0.89637152
#> 56  056 -0.410452325 -0.53508134
#> 57  057 -0.623623214 -0.53838913
#> 58  058 -0.027188146  0.74102314
#> 59  059 -2.683469135 -0.32473123
#> 60  060 -0.038088087 -1.00687622
#> 61  061 -0.286178183 -0.28279416
#> 62  062  1.030994856  2.73081306
#> 63  063 -1.348562051 -0.53061185
#> 64  064  1.482203591  1.17380408
#> 65  065  0.504399705  0.70054029
#> 66  066 -0.639753344 -1.27665050
#> 67  067  0.988084908  0.08345196
#> 68  068  0.337995342  0.76738667
#> 69  069 -1.373393895 -0.73442190
#> 70  070 -0.326123194  0.57664308
#> 71  071  0.826504490  0.01560202
#> 72  072  1.337940124  0.08132418
#> 73  073  1.109967526  0.13751503
#> 74  074  0.234177221 -0.84371277
#> 75  075 -0.938435775 -0.99845648
#> 76  076  0.971892254 -0.49959492
#> 77  077 -1.863883399 -0.43632272
#> 78  078  0.224420054  0.85696948
#> 79  079  0.829167154 -0.34717744
#> 80  080 -2.762447909 -0.38707498
#> 81  081  0.447074585 -0.12091882
#> 82  082 -0.928266850 -1.39824797
#> 83  083 -1.084771863  0.50439212
#> 84  084 -0.789096821 -1.05120166
#> 85  085  0.015296929 -1.99088207
#> 86  086 -0.589714683 -0.87989323
#> 87  087 -0.050514139  0.20414271
#> 88  088  0.520234109  0.57572637
#> 89  089  0.541507778 -0.72124394
#> 90  090 -0.682485796 -0.41566470
#> 91  091 -0.541173498 -0.88258826
#> 92  092  0.841234496  2.92571307
#> 93  093  0.705879465  0.30947673
#> 94  094  0.923883371  1.18266583
#> 95  095 -1.797867059 -0.62627615
#> 96  096 -1.224833540  0.35263881
#> 97  097 -0.152752448  0.53906841
#> 98  098 -1.890752007 -0.90063166
#> 99  099 -1.373815886 -1.52238595
#> 100 100  0.014923251  1.22415735