Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4976115
#> Y_2 0.4976115 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4676517 -0.7412034
#> 2 -1.2504100 -1.6246978
#> 3  1.9188541  1.0743809
#> 4  2.0874785  1.3419528
#> 5  1.1617143  1.0709418

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.51964668  0.330188330
#> 2   002 -0.95008570 -1.028594676
#> 3   003  0.03130784 -0.159769616
#> 4   004 -0.31881652  0.453489862
#> 5   005  1.15618276  1.434493501
#> 6   006 -1.54732167 -2.183163048
#> 7   007  0.51293061  0.260940596
#> 8   008  0.35086496  1.841469231
#> 9   009 -0.35161321  0.237312377
#> 10  010 -0.04816941 -0.967330070
#> 11  011  0.10171941  1.536149443
#> 12  012 -0.05593297  0.227374674
#> 13  013 -0.80157547 -0.631438260
#> 14  014  0.16615920  1.131523619
#> 15  015 -0.30135430  1.544903089
#> 16  016  0.13740559 -1.805478643
#> 17  017  1.16393605  1.021105895
#> 18  018  2.00831850  1.208889543
#> 19  019  0.21104045  0.007442123
#> 20  020 -0.59367509 -0.937499862
#> 21  021  0.93834780 -0.010970269
#> 22  022  0.44510359 -0.036850787
#> 23  023  1.39700117 -0.716613592
#> 24  024  1.41161549  0.800343477
#> 25  025 -0.03394507 -0.044414215
#> 26  026 -0.98920250  0.624393998
#> 27  027  2.08157531 -0.384502182
#> 28  028 -0.49537728 -0.654334790
#> 29  029  0.84030734  0.824863436
#> 30  030  1.40738051  0.343449546
#> 31  031 -0.42403292 -0.505000302
#> 32  032 -0.71782139  0.989964462
#> 33  033 -0.65521020 -0.037292712
#> 34  034  0.49061067  2.057616628
#> 35  035  1.56710557  1.828911024
#> 36  036  0.33578161 -1.491129988
#> 37  037  0.09451628 -0.258874933
#> 38  038 -0.17127895  1.020969208
#> 39  039 -0.36094202  0.622664884
#> 40  040 -0.77979317 -0.573045768
#> 41  041  0.10129419  0.846547568
#> 42  042  0.85623580 -0.012138524
#> 43  043  1.43839452  0.961715369
#> 44  044 -0.28251118 -0.070463306
#> 45  045  0.40225072 -0.377891051
#> 46  046  0.98264612  1.369292330
#> 47  047 -0.09461835 -0.361860835
#> 48  048  0.81801963  0.831636301
#> 49  049  0.45635921  1.049625479
#> 50  050 -1.02426096 -0.593815781
#> 51  051 -0.28557606 -0.982934918
#> 52  052  1.34129320  1.619945297
#> 53  053 -1.23963830  0.623560022
#> 54  054 -0.64319764  1.320933332
#> 55  055 -0.53609511  0.850075528
#> 56  056  2.84068483  1.460186504
#> 57  057  1.08161111 -0.303113254
#> 58  058  2.18932137  1.540001274
#> 59  059  0.27501663  0.166958911
#> 60  060 -0.07325028 -0.523015928
#> 61  061 -0.30622764 -0.829708808
#> 62  062 -0.41967141  1.081975079
#> 63  063 -1.04774094  0.026916717
#> 64  064  1.43285122  0.762130417
#> 65  065 -1.11241341  0.007773616
#> 66  066  0.18604869 -1.313114564
#> 67  067 -0.30924827 -1.579081685
#> 68  068 -0.65366674  1.062754532
#> 69  069 -0.03966878 -0.121282872
#> 70  070 -1.32659350 -0.921675710
#> 71  071 -1.22898808  0.374808505
#> 72  072  0.42156414  0.537866543
#> 73  073 -0.29606997 -0.108217097
#> 74  074  0.41624061 -0.599913612
#> 75  075 -1.25239816 -0.161337800
#> 76  076  0.58481590  0.268212779
#> 77  077 -0.28867779  1.398560097
#> 78  078  0.27279533 -0.572038333
#> 79  079 -0.42038602  0.310362839
#> 80  080 -0.16613307 -2.103061149
#> 81  081 -0.30037192 -2.048634312
#> 82  082  0.78998495  1.574110712
#> 83  083  0.37789244 -0.362045484
#> 84  084 -0.98301317 -0.116047277
#> 85  085  0.44608497  1.390439909
#> 86  086  0.93727819  0.334329870
#> 87  087  0.56415944  0.007094732
#> 88  088 -0.74760146 -0.888608916
#> 89  089  0.20606918  1.256949004
#> 90  090 -0.94557811 -0.744251751
#> 91  091 -1.07989827 -0.471508519
#> 92  092  0.00266030  1.101225280
#> 93  093 -0.49213112 -0.146691611
#> 94  094  0.03326728  0.358911417
#> 95  095 -0.96052676 -0.636567432
#> 96  096  0.99632662  0.099709992
#> 97  097  0.90420784  1.394261150
#> 98  098 -0.85692243  0.037932645
#> 99  099 -0.03238106 -0.302722012
#> 100 100 -0.10725945 -0.748641626

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.711777104 -2.18579208
#> 2   002  0.307808784  0.40808375
#> 3   003  1.135428838  1.39558471
#> 4   004 -0.085875861  0.50384425
#> 5   005 -0.369562687 -0.37250520
#> 6   006 -0.603295874 -0.56418658
#> 7   007 -1.214737153  0.93515247
#> 8   008  1.648369441  0.00688533
#> 9   009 -0.309356170 -0.92800862
#> 10  010  0.340311681  0.40418540
#> 11  011  2.259562729  2.30071873
#> 12  012  0.986128852  1.67197925
#> 13  013  2.085254949  0.86879442
#> 14  014  0.846839155  1.93136239
#> 15  015 -0.229695712 -1.22396098
#> 16  016  0.401403231  0.82700987
#> 17  017  0.142224289 -2.23992413
#> 18  018 -0.136102264  0.07253737
#> 19  019 -0.023800260  0.81436518
#> 20  020 -0.393292952  0.48974744
#> 21  021  2.044170801  1.80410797
#> 22  022  1.222507575 -1.23140589
#> 23  023 -0.421773367  1.05440154
#> 24  024  0.758267404 -0.41802177
#> 25  025 -0.568514195  0.02474988
#> 26  026 -1.722359382  0.05397272
#> 27  027  1.077446270  1.33401934
#> 28  028 -1.980712503 -1.11230112
#> 29  029 -0.141905346 -0.50270223
#> 30  030  0.269236067  0.20743857
#> 31  031 -0.174172108  0.45482215
#> 32  032  1.459407562  2.07892549
#> 33  033 -0.765090487 -0.77143046
#> 34  034 -0.304925463  0.12763241
#> 35  035  0.728907581  1.42946662
#> 36  036  1.197624878  1.01860996
#> 37  037  0.685281620 -0.21287591
#> 38  038 -2.087513178 -0.45048419
#> 39  039 -1.806397593 -0.83224356
#> 40  040 -2.199550054 -2.89988663
#> 41  041  0.079705500 -0.50167468
#> 42  042  1.881239578  1.83558350
#> 43  043  0.461688471  0.38476183
#> 44  044 -1.286601801  0.59182834
#> 45  045  0.170307305  1.51760117
#> 46  046 -1.049936148 -0.74671183
#> 47  047  1.081043090 -0.16464141
#> 48  048 -2.639913131 -1.20188109
#> 49  049 -1.275156591 -0.67255101
#> 50  050 -0.674983561  0.26040260
#> 51  051  1.092656976  1.54847705
#> 52  052 -1.589310750 -0.17385143
#> 53  053  1.007560789 -0.11046295
#> 54  054  0.712696659  1.70941096
#> 55  055  2.365305962  2.47005533
#> 56  056  0.644738561 -1.43372546
#> 57  057  0.456206455  0.07948848
#> 58  058 -1.872100282 -0.75272415
#> 59  059  0.699511371  0.57649791
#> 60  060 -1.669112413 -1.93303037
#> 61  061  0.619519623 -0.31126076
#> 62  062  1.288641209  0.99244216
#> 63  063 -0.185284663  0.03118765
#> 64  064 -0.527165568 -0.69576740
#> 65  065  2.157177246  0.52996741
#> 66  066  0.983737900  1.14635993
#> 67  067 -0.371544952 -0.07078699
#> 68  068  0.845827052 -0.37625607
#> 69  069 -1.046388901 -0.70519466
#> 70  070  0.120410734 -0.73266371
#> 71  071  0.416476090  1.26869595
#> 72  072  0.139117673 -0.55243287
#> 73  073  1.803337123  1.64904756
#> 74  074  0.316610764 -0.18295297
#> 75  075 -0.951495545  1.35680706
#> 76  076  0.742245897  0.05050376
#> 77  077  0.375027812  1.25013930
#> 78  078  0.740108903  1.02801113
#> 79  079  0.415865006  0.60375020
#> 80  080 -1.559777129 -0.63731338
#> 81  081  0.693809470 -0.07058866
#> 82  082  0.064613756 -0.96269128
#> 83  083 -0.238879212 -0.25184661
#> 84  084  1.432639321  1.00348687
#> 85  085 -0.780241585  0.45273696
#> 86  086 -1.053103620 -0.06493115
#> 87  087  0.110013110  0.93021200
#> 88  088  0.755672026  0.21245425
#> 89  089 -0.456760505  0.02743550
#> 90  090 -0.830429479 -1.35769368
#> 91  091  0.005117279  0.97422899
#> 92  092 -0.694889373 -1.31619482
#> 93  093  0.897585934  0.44749622
#> 94  094  1.254463533  0.28159615
#> 95  095  0.822349171  1.05867325
#> 96  096 -1.682467173 -1.17069069
#> 97  097 -0.455921911  1.19062275
#> 98  098  0.708950794 -0.38698765
#> 99  099  0.281186435 -0.38545106
#> 100 100 -0.679787027 -0.13832647