Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5184516
#> Y_2 0.5184516 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  2.2407366  1.2143024
#> 2 -0.1411780  0.1381138
#> 3  1.3480169 -0.7264422
#> 4  0.2807788  1.6225408
#> 5  1.1401776 -0.6666642

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -2.40193792 -2.418682323
#> 2   002  1.23589336  0.425171337
#> 3   003  1.67516818  0.832269341
#> 4   004 -0.17576493 -0.506687552
#> 5   005 -0.94230682 -0.115738416
#> 6   006  0.66360043  2.343269999
#> 7   007 -0.03646479  0.006874699
#> 8   008 -0.07455378  1.330690781
#> 9   009  2.51773140  1.827280823
#> 10  010 -0.77847459 -1.211766220
#> 11  011 -1.10581560 -1.484066296
#> 12  012  0.23118865  0.632480464
#> 13  013  0.48839898  0.120786493
#> 14  014 -1.93709478 -2.348461028
#> 15  015  1.21918775  0.660606137
#> 16  016 -0.74734371 -1.027198637
#> 17  017  0.17684407  0.724768114
#> 18  018 -1.84890537 -3.512407403
#> 19  019  0.35003973  1.640845700
#> 20  020 -0.53646099  0.122688593
#> 21  021 -1.10786808 -0.407506398
#> 22  022  1.47080003  0.060613832
#> 23  023 -0.12024793  1.519657908
#> 24  024  0.22792914 -0.391303358
#> 25  025  1.04205905  0.147549793
#> 26  026 -1.61981416 -1.364473368
#> 27  027 -0.06296767 -0.904052106
#> 28  028  0.64510656  2.118829431
#> 29  029  1.86406062  1.140255026
#> 30  030  0.65653081 -0.176080946
#> 31  031 -0.33476102 -0.199633125
#> 32  032 -0.29759594  0.910802231
#> 33  033 -1.80030608 -0.206807076
#> 34  034  1.25843793  0.842715209
#> 35  035 -1.23238524 -0.903481561
#> 36  036 -1.33639360  0.025253105
#> 37  037 -1.10740175 -1.822962320
#> 38  038  0.88003241  0.133210988
#> 39  039 -0.24147830 -0.503426955
#> 40  040 -1.13065693  0.039045698
#> 41  041 -0.83720680 -0.928558354
#> 42  042 -0.42845801 -0.101346601
#> 43  043 -0.99777979  0.746099706
#> 44  044  0.14149132  1.269997997
#> 45  045  0.83090326  0.134959085
#> 46  046 -0.28461347  0.437298265
#> 47  047 -0.54216490 -0.670781156
#> 48  048 -1.32472893 -1.196901823
#> 49  049  0.07489984  0.051374509
#> 50  050 -0.07729959  0.362749742
#> 51  051 -1.23264480 -0.230605375
#> 52  052 -0.71439795  0.103952090
#> 53  053 -0.12252547  0.550210620
#> 54  054  0.42283367  1.100262529
#> 55  055  0.43293675  0.405572689
#> 56  056  1.85910521  1.778160963
#> 57  057  0.50670251  0.808525710
#> 58  058  0.27111992  0.541296697
#> 59  059 -0.01632957  0.270462964
#> 60  060  0.25793434  0.292506138
#> 61  061  0.72548720 -0.341275494
#> 62  062  0.29547917  1.198122469
#> 63  063  0.56109404  1.681304453
#> 64  064 -0.98317299  0.390177435
#> 65  065 -1.42865001 -0.273164872
#> 66  066  1.42307840  0.282779026
#> 67  067 -0.10641459 -1.400524225
#> 68  068 -1.12680280 -1.073201763
#> 69  069  0.04948889  0.081486603
#> 70  070  0.26101157 -0.780756501
#> 71  071 -0.63774092 -1.249914681
#> 72  072 -0.67607149  0.036589203
#> 73  073 -2.15935272 -1.460189252
#> 74  074 -0.42926203  0.134543778
#> 75  075 -0.30635645 -0.080552135
#> 76  076  0.04410156  0.270774743
#> 77  077 -0.37969526 -0.375416149
#> 78  078  0.02725168  1.616000113
#> 79  079 -0.49173383  0.699580699
#> 80  080  0.25698195 -0.004931074
#> 81  081  1.75893524  0.870555771
#> 82  082 -0.43712353 -0.093316031
#> 83  083 -0.43278676 -0.980309443
#> 84  084 -1.43324535 -1.119138564
#> 85  085  0.71836024  1.724462275
#> 86  086  2.45327730  1.286627506
#> 87  087 -0.49751508 -2.025116287
#> 88  088 -0.84100727  0.466361422
#> 89  089 -0.35198514 -0.079936490
#> 90  090 -1.83047383 -0.236354320
#> 91  091  0.45753925 -0.384387665
#> 92  092  1.02642071  0.752136955
#> 93  093 -1.38510840  0.593973054
#> 94  094  1.17182294  0.763687866
#> 95  095  0.75972819  1.113682437
#> 96  096 -0.03763085  0.311572608
#> 97  097 -1.38301016 -1.739221384
#> 98  098  0.11911521  1.139915120
#> 99  099  0.14660171 -0.738581821
#> 100 100  0.01937873  1.076139696

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1           Y.2
#> 1   001  0.8206496629 -0.7133458520
#> 2   002 -0.0623236541 -0.1530010928
#> 3   003  0.2369150008 -1.4786273415
#> 4   004  0.1398790000 -0.7140660820
#> 5   005  0.1578867461 -0.1231577888
#> 6   006 -0.4794251230 -0.8501658304
#> 7   007 -0.4385171248 -0.9649879202
#> 8   008 -0.4075240727  0.9163572602
#> 9   009  0.5091033579  0.9739943237
#> 10  010  2.0117234065  0.7237244697
#> 11  011 -1.1393627687 -0.3637803117
#> 12  012  0.2852253019 -0.2021108415
#> 13  013  0.6397498742 -0.9709266036
#> 14  014  1.7600240145  1.0667758060
#> 15  015  0.1955875468 -0.5631169102
#> 16  016  1.2554780872  0.2976359992
#> 17  017  1.5844892240  0.9791676136
#> 18  018 -2.7523731799 -1.2787801037
#> 19  019  1.8493561062 -0.1635755948
#> 20  020 -0.7987774820 -0.3412810877
#> 21  021  0.6102535228  0.4045178819
#> 22  022 -0.9722284732  0.6685396428
#> 23  023 -0.9753020937 -0.4063159319
#> 24  024  0.2820681429 -0.0136604287
#> 25  025 -0.5155321200  1.2500049717
#> 26  026 -0.2071391580 -0.4271022737
#> 27  027  0.8180985584  1.0866092275
#> 28  028 -0.9312610988 -0.3352226012
#> 29  029 -0.7248464068 -0.5035921580
#> 30  030  0.6013689223 -0.6199822919
#> 31  031  0.4577530155  0.4862483365
#> 32  032  2.2405251137  0.0540779907
#> 33  033  0.4258441761  0.3197656909
#> 34  034 -1.0018184908  0.9965218674
#> 35  035  0.9912733240  1.5581235784
#> 36  036 -1.1734634558 -1.2091455300
#> 37  037  0.4927450453  1.4153206504
#> 38  038 -1.0742569802 -0.8095406830
#> 39  039  0.4186426534 -1.0953763497
#> 40  040 -0.7481658679  1.1244847050
#> 41  041  1.0155496353  0.9525932467
#> 42  042 -0.1749831149 -1.1990271490
#> 43  043 -1.3845043448 -2.1876685461
#> 44  044  0.0030320697 -1.2317287186
#> 45  045 -0.2009515356  0.9420467072
#> 46  046 -0.5884660434 -0.7489890831
#> 47  047 -1.1852021491  0.5983027994
#> 48  048  0.3094832186 -0.5593819866
#> 49  049  0.7890722773  1.4738643943
#> 50  050 -0.5586564984 -0.1433259369
#> 51  051 -1.0801076815 -0.7392460367
#> 52  052 -0.6842824222 -0.4620006013
#> 53  053 -0.3851080474  0.2670196068
#> 54  054 -1.3332403882 -1.8677351655
#> 55  055 -0.8861670707  0.7098936536
#> 56  056 -1.7166875277 -2.3037284346
#> 57  057  0.4505808957  0.6343949352
#> 58  058 -0.0765702209  0.7492256448
#> 59  059 -0.3566672632  0.6935710877
#> 60  060 -0.0858740847 -0.7871692969
#> 61  061  1.1723734346  0.4876245231
#> 62  062  0.2140011424  0.0005709324
#> 63  063 -1.1462456082 -1.4142811867
#> 64  064 -0.5369281457 -0.0166239442
#> 65  065 -0.8378174152 -0.5225229642
#> 66  066 -0.0001741209 -0.4277607825
#> 67  067 -0.2927637120 -1.8536935260
#> 68  068 -2.2572122165 -0.6084724716
#> 69  069  1.4202024590  0.6930744121
#> 70  070 -1.9606988720 -0.0971398187
#> 71  071  0.6834184333  0.6237280154
#> 72  072 -0.8484296578 -0.5736512951
#> 73  073  0.7002731792  0.8170718903
#> 74  074 -0.2737273208 -0.5559167510
#> 75  075 -2.3995445432  0.6844226845
#> 76  076 -0.9341663097  0.2306599840
#> 77  077  0.7832051603  0.3409349768
#> 78  078 -0.5861993139 -1.5215252116
#> 79  079 -0.3269516353  0.0904597379
#> 80  080 -0.1136182455 -0.7746660087
#> 81  081 -0.0748597640 -0.9866622477
#> 82  082 -1.9644608831 -0.0662497700
#> 83  083 -0.7077063289 -0.8054820631
#> 84  084  0.4309130893  1.2296714099
#> 85  085 -0.5225854107  0.8143704259
#> 86  086 -0.5803424680  1.4532749922
#> 87  087  0.3813325588  2.0580017688
#> 88  088  0.8704074177  0.7602723248
#> 89  089  0.0760541251  1.2123616383
#> 90  090  0.3629039762 -0.6988923466
#> 91  091  0.0990514786  0.4867254837
#> 92  092 -0.6366180337  0.6030440474
#> 93  093 -0.5180633449 -0.0951715295
#> 94  094 -0.3612408013 -0.4254270768
#> 95  095  0.6100526779  1.6205300291
#> 96  096  0.7928075885  0.5333963376
#> 97  097  0.8485829680  1.4350777654
#> 98  098 -0.8055475593  0.5555836551
#> 99  099 -0.1740285466 -1.2734510293
#> 100 100  0.6681534076 -0.0571558442