Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.496303
#> Y_2 0.496303 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.1385506  2.2405663
#> 2 -0.3725024 -0.3439459
#> 3  1.5149711  0.1499078
#> 4 -0.6517449  0.3080084
#> 5 -0.8042954  0.4996406

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.909051047 -1.02237945
#> 2   002  0.312722099 -0.30250578
#> 3   003  0.699806004  0.99562832
#> 4   004  0.596798190  0.55619765
#> 5   005 -0.153028898 -0.33181917
#> 6   006  0.309201254  0.00669120
#> 7   007  1.776359281  1.21396307
#> 8   008  1.083588844  0.66324885
#> 9   009 -1.245992438  0.38453905
#> 10  010 -0.472374569 -0.74409557
#> 11  011  2.841076570  0.58398818
#> 12  012  1.686053322 -0.78652075
#> 13  013 -0.292013186  1.54583663
#> 14  014 -0.344343607 -0.85974982
#> 15  015 -0.524904212 -1.01388220
#> 16  016 -0.368896079 -0.94825543
#> 17  017  0.153100698  0.06981221
#> 18  018  0.029199266 -0.09790244
#> 19  019  1.041258168  0.40036398
#> 20  020  2.486231998  1.66197096
#> 21  021  0.024922344 -0.82429386
#> 22  022  0.077204544 -0.26590116
#> 23  023 -0.551855392 -0.01580600
#> 24  024  0.715181801 -0.06080226
#> 25  025  1.102506002  0.63665549
#> 26  026 -1.848471160 -0.92206980
#> 27  027  0.932882768  0.40198730
#> 28  028  0.213970472  0.27530389
#> 29  029  1.689557000  1.47281178
#> 30  030  1.357472040  2.27902022
#> 31  031  0.650293236  1.47811931
#> 32  032  0.537729035 -0.63782037
#> 33  033 -1.024318651 -0.97843890
#> 34  034  0.586166319  0.08155733
#> 35  035  1.264693822  2.02482236
#> 36  036  2.158395415  2.68437451
#> 37  037  1.075356926  1.86675890
#> 38  038  2.063051269  1.48053442
#> 39  039  0.041868381  1.33637640
#> 40  040  0.755425039 -0.84093014
#> 41  041  2.297941292  0.03960070
#> 42  042  1.136454355  0.29777574
#> 43  043 -0.989450827  0.60614121
#> 44  044  0.054501473 -0.36539891
#> 45  045  0.548235229  1.59623842
#> 46  046 -1.038231215 -0.76698158
#> 47  047  0.946798160 -0.27794502
#> 48  048  0.329049394  1.58420085
#> 49  049 -0.478988487 -2.36406004
#> 50  050  1.307200150  0.28280589
#> 51  051 -0.267893330 -0.89162907
#> 52  052 -0.300681796 -0.25315965
#> 53  053  1.257743730  0.89727076
#> 54  054 -0.002241241  0.64003254
#> 55  055 -0.703438424 -0.11848550
#> 56  056 -1.252084321 -0.91321190
#> 57  057 -0.791104159 -2.28967440
#> 58  058 -1.518471416 -0.46898611
#> 59  059  1.418846907  0.23752956
#> 60  060  0.977681405  0.11247881
#> 61  061  0.660212412  0.04070071
#> 62  062 -0.197129127 -0.65209968
#> 63  063  0.355893560  1.01412050
#> 64  064  0.403784655  1.95423083
#> 65  065 -0.225254392 -0.31917786
#> 66  066 -0.148646137  0.57461259
#> 67  067  1.159761141  0.01476683
#> 68  068  0.508506720  0.51061690
#> 69  069 -1.096689884 -0.30528958
#> 70  070  0.032258452  0.84074754
#> 71  071 -0.211710557  1.67703036
#> 72  072  0.224953112 -0.13764350
#> 73  073  0.414933511  0.95346407
#> 74  074  0.916643632  0.74435194
#> 75  075 -1.216483185 -1.90725513
#> 76  076 -0.087224235  0.70292326
#> 77  077 -1.193136364 -0.70962318
#> 78  078  0.063245776 -0.64211200
#> 79  079  0.765943967  0.89468968
#> 80  080 -1.085756629 -0.73095914
#> 81  081 -1.309926697 -0.21053284
#> 82  082 -1.548464176 -1.52314179
#> 83  083 -1.256650050 -1.33253219
#> 84  084  0.133922606  0.15983611
#> 85  085  0.855493627 -0.08209776
#> 86  086  0.511987943 -0.16561279
#> 87  087 -0.733554653  0.30196360
#> 88  088 -0.136191523  0.03412969
#> 89  089  0.398436502 -0.43678940
#> 90  090 -0.488282334 -1.23332115
#> 91  091  0.789822911 -1.16437895
#> 92  092 -0.611975319 -0.70426577
#> 93  093 -0.465674742  0.28507864
#> 94  094 -0.557946504 -1.64737546
#> 95  095  0.444901403 -0.35632130
#> 96  096 -1.563780648  0.31460401
#> 97  097  0.921366609 -0.22421237
#> 98  098  1.278860323  0.08021920
#> 99  099  0.661728974 -0.98710633
#> 100 100  1.496921700  0.83325752

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.957527723  1.367110770
#> 2   002 -0.780764159 -1.045371955
#> 3   003  0.284641657  1.465028359
#> 4   004  1.130827372 -0.552362686
#> 5   005  1.324233829  1.042811084
#> 6   006  0.622320281 -0.311767601
#> 7   007 -0.601713056 -0.826878289
#> 8   008  0.281087276 -1.587054561
#> 9   009 -0.188118333  1.275926033
#> 10  010  0.883521595  0.567932526
#> 11  011  1.501437192  0.473536109
#> 12  012  0.880236993  1.280545630
#> 13  013 -0.591738851 -1.359001874
#> 14  014 -0.203407831  0.204278503
#> 15  015 -0.471155330  0.027519720
#> 16  016 -1.264419465  0.493319922
#> 17  017  0.702317535 -0.715752260
#> 18  018 -0.454281841  0.031805748
#> 19  019 -1.753186280 -1.143619934
#> 20  020 -0.008477976 -2.242291676
#> 21  021  0.085996857  1.509459603
#> 22  022  0.631619016  0.975966603
#> 23  023  1.003210595  1.450126192
#> 24  024  0.872016386  1.378552407
#> 25  025  1.138314961  1.049394140
#> 26  026 -0.815478380 -0.857563645
#> 27  027  1.273335157 -0.005702675
#> 28  028  0.457124988  0.264847903
#> 29  029  0.177534107  1.144582925
#> 30  030  2.295584072  1.326409331
#> 31  031 -1.122079627  0.047838265
#> 32  032  0.874829943  0.774758563
#> 33  033 -0.094915149 -1.664218488
#> 34  034  0.337115935 -0.782548611
#> 35  035 -0.002820388  1.541327981
#> 36  036  0.181670739  1.115392025
#> 37  037 -1.829407634 -1.384496477
#> 38  038  1.300769943  0.052743622
#> 39  039  1.928132524  1.125433186
#> 40  040 -1.197010508 -0.683648100
#> 41  041  0.753990733  0.748252182
#> 42  042 -1.271532677 -0.223210806
#> 43  043 -0.275025152  0.633298066
#> 44  044  0.309439459  0.854898708
#> 45  045 -1.653705281 -0.504972747
#> 46  046 -1.347259695 -2.007899303
#> 47  047 -0.451195105 -0.080074083
#> 48  048 -0.735754114 -2.814286754
#> 49  049 -0.574606141 -1.130739991
#> 50  050  0.523774263  0.663711101
#> 51  051  0.504481993 -0.313874106
#> 52  052  0.854660128  1.038606579
#> 53  053  1.091877227 -0.739449843
#> 54  054 -0.523614004 -1.626626227
#> 55  055  1.225899229 -0.035924028
#> 56  056 -0.196110572  0.021289963
#> 57  057  0.522118606  1.150513297
#> 58  058 -1.627721382 -0.438087036
#> 59  059 -0.752949866  0.155298161
#> 60  060  0.070364620 -0.208458670
#> 61  061  0.442290374  0.220746332
#> 62  062  2.244056400  1.733119802
#> 63  063  0.141218577  1.051578329
#> 64  064 -0.935051016  1.643072655
#> 65  065  1.998001285  0.405842101
#> 66  066 -0.292714669 -0.193149006
#> 67  067  0.276538351 -1.687590320
#> 68  068  0.220145742 -0.816991168
#> 69  069  1.375955048  1.040812711
#> 70  070 -0.765116147  0.768362134
#> 71  071  0.560938690  0.003665458
#> 72  072 -0.663639342  1.067660518
#> 73  073 -0.419377776 -0.173011903
#> 74  074 -0.126323523  1.178629886
#> 75  075  1.857538394 -0.669414155
#> 76  076  0.918576507  1.546119403
#> 77  077 -1.376393302 -1.762088206
#> 78  078 -1.274361440 -1.323087782
#> 79  079 -0.798919304 -0.039450826
#> 80  080  0.479412544 -0.249219196
#> 81  081 -0.713386642 -1.362471805
#> 82  082 -0.030057285  1.428548457
#> 83  083  1.119965046  0.396223226
#> 84  084 -0.425769623 -1.710527655
#> 85  085 -0.629210876 -1.328504259
#> 86  086  1.815372006  1.900190560
#> 87  087 -0.096595636  0.894139919
#> 88  088  0.552239080  1.412780319
#> 89  089  0.867862750  0.387141013
#> 90  090 -0.054837258 -1.809092404
#> 91  091  1.286058551  0.988354669
#> 92  092 -1.845308262 -1.571938349
#> 93  093 -0.041095332 -0.887855807
#> 94  094 -0.207125248  0.518455445
#> 95  095 -0.957957783  0.023331073
#> 96  096 -0.185342409 -1.506758673
#> 97  097 -0.143942231  0.175653752
#> 98  098 -0.884074768 -0.003212592
#> 99  099 -0.695512817  0.282310472
#> 100 100 -2.519695480 -1.210460794