Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.470661
#> Y_2 0.470661 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.9248455  1.6604630
#> 2 -1.4575707  0.6250634
#> 3 -0.4377923 -3.0360510
#> 4 -0.5373098  0.2432571
#> 5 -0.6096128 -0.8979677

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.528965226 -1.201211009
#> 2   002 -0.627785821 -0.711300777
#> 3   003  0.821582556  0.912375567
#> 4   004  1.068684739  1.065072059
#> 5   005  0.267702142  1.015584419
#> 6   006  0.482739090  0.582434250
#> 7   007  1.315160579  1.382120850
#> 8   008 -0.826639540 -0.515725555
#> 9   009 -0.388012521  0.058105692
#> 10  010 -1.267305974  0.855470533
#> 11  011  0.608329546  0.628211389
#> 12  012  1.880948654  2.293751022
#> 13  013 -0.276572244 -0.128939641
#> 14  014  0.994870795  0.102323696
#> 15  015 -0.710510957 -1.501062121
#> 16  016  0.390847822  0.671227116
#> 17  017 -0.350283519  1.180492046
#> 18  018  0.543449265  1.145478784
#> 19  019  0.112083932  0.776256008
#> 20  020 -0.212665213 -0.082640508
#> 21  021 -1.668187763 -2.406665662
#> 22  022 -1.346069568 -1.448428354
#> 23  023 -0.841944138 -0.988556127
#> 24  024  0.885952666 -0.735577502
#> 25  025 -0.367599813 -0.229924979
#> 26  026 -0.006756171  0.107246073
#> 27  027  2.299428721  1.500813357
#> 28  028  1.417730841  2.257177547
#> 29  029  1.440522983 -0.315064555
#> 30  030 -0.107823888 -0.122612561
#> 31  031 -0.113538143 -0.399593053
#> 32  032 -0.420666996  0.197020836
#> 33  033 -0.327362813  0.556062395
#> 34  034 -0.107871035  0.014214109
#> 35  035 -0.213634213 -0.327263049
#> 36  036 -0.865616669  0.188470111
#> 37  037 -1.361563134  0.584227807
#> 38  038 -0.983640518  1.040686390
#> 39  039 -0.083755951 -1.904743047
#> 40  040 -1.472617892 -1.618651069
#> 41  041 -1.142045417 -0.684190860
#> 42  042  0.233471272  1.927032069
#> 43  043  1.134888605  1.891842555
#> 44  044  0.284113395 -0.903137154
#> 45  045  0.713463403  0.901637945
#> 46  046  0.568221348  2.223513271
#> 47  047 -0.178204412 -0.385965497
#> 48  048  0.932567269  0.762045995
#> 49  049 -0.189721560 -0.894975346
#> 50  050 -0.112893316 -0.314560488
#> 51  051  0.256133467 -2.542244482
#> 52  052 -1.370104258 -0.506347200
#> 53  053 -2.494537841 -1.764618846
#> 54  054 -0.079325070 -0.283620611
#> 55  055  1.284546957  0.535099025
#> 56  056  0.623874822  0.332118213
#> 57  057 -0.464454156  0.900476967
#> 58  058  0.150457565  0.832161214
#> 59  059 -1.225104222 -0.836314561
#> 60  060  2.778658907  1.223210797
#> 61  061 -0.428562253  0.086074170
#> 62  062  1.673901409  0.678395749
#> 63  063 -0.122544972  1.033510379
#> 64  064 -0.364892005  0.822140279
#> 65  065  0.486597267  0.182003251
#> 66  066  0.578670904  0.678440833
#> 67  067  0.780728816  0.761047874
#> 68  068 -0.795999085 -0.405421794
#> 69  069 -1.429732375 -1.031887721
#> 70  070  0.708548871  0.907315188
#> 71  071  1.000686505  0.222148479
#> 72  072 -1.460206301  0.002241018
#> 73  073  1.165332701  1.550240267
#> 74  074 -0.410426433 -0.143852867
#> 75  075  0.062804964 -0.060597792
#> 76  076 -0.687730137  0.333736919
#> 77  077  1.943385384  1.546695433
#> 78  078  0.610265694  0.760765792
#> 79  079 -0.311861471  0.302803868
#> 80  080 -0.772296589 -1.333034115
#> 81  081  0.893745755 -0.371805887
#> 82  082  0.569498828  0.051457474
#> 83  083  1.256469469 -0.515065265
#> 84  084 -0.488481553 -0.525946189
#> 85  085  1.648696974  0.400563475
#> 86  086 -1.289111251  0.316077775
#> 87  087  0.879726935 -0.345174138
#> 88  088 -0.054577538 -0.869639920
#> 89  089  0.240558719 -0.053243101
#> 90  090  0.969667209  1.324540368
#> 91  091  0.309823423 -0.322087968
#> 92  092  0.704411489  0.619343733
#> 93  093  0.820578466  0.298421746
#> 94  094 -0.328684365 -1.090123449
#> 95  095 -0.105751216  0.509675773
#> 96  096 -1.243424282 -0.307852663
#> 97  097  0.429478927  0.592346603
#> 98  098  0.536581286  0.419300834
#> 99  099  0.441717606  0.538112318
#> 100 100  0.355916495 -1.300474713

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.08037244 -1.17358468
#> 2   002 -0.34061175  0.14161096
#> 3   003 -0.75531076  0.78745614
#> 4   004  0.54008890  0.29438486
#> 5   005 -1.82782655 -1.13619882
#> 6   006  0.13502725  0.16632598
#> 7   007 -0.18003133  1.64998166
#> 8   008  0.29364513 -0.24597058
#> 9   009  1.08001703  1.60861458
#> 10  010 -0.51335103 -0.20096101
#> 11  011 -0.33511138  0.61065066
#> 12  012 -0.13600959  0.07160176
#> 13  013  0.64586536 -0.79108175
#> 14  014 -0.15025158  0.75681666
#> 15  015  0.22130792 -0.72571412
#> 16  016 -0.10412528  0.75338954
#> 17  017 -0.76318936  2.00442606
#> 18  018  1.21555873  0.35271615
#> 19  019  0.58572707 -0.20147254
#> 20  020  0.17957893  0.44333222
#> 21  021  0.28971658  0.30594455
#> 22  022 -0.76227089 -1.10671152
#> 23  023 -2.43221610 -1.42650854
#> 24  024 -0.27883392  0.61375647
#> 25  025  0.37207525  0.07300824
#> 26  026  0.20594493 -0.86490109
#> 27  027 -0.72305296  0.98829362
#> 28  028  0.26894945  0.12583073
#> 29  029  0.19253488 -0.12900442
#> 30  030 -0.57088748  0.56440447
#> 31  031 -0.39628766 -0.67437401
#> 32  032  1.35040654 -0.10549830
#> 33  033 -1.49931567 -2.55254596
#> 34  034  0.84721990 -0.14790951
#> 35  035  0.85089368  0.69223794
#> 36  036 -0.25949157 -0.70975611
#> 37  037 -1.70151597 -1.53299055
#> 38  038 -0.30564826 -0.86705531
#> 39  039 -0.28280612  0.10474351
#> 40  040  0.58026009  0.35526585
#> 41  041 -1.31814391 -0.93511694
#> 42  042 -0.75890653  0.32340419
#> 43  043 -2.10876500 -0.66057903
#> 44  044  0.11153560  1.14088503
#> 45  045  0.15136139 -1.42713953
#> 46  046  0.48449175 -0.58922300
#> 47  047 -1.10476989 -2.02855590
#> 48  048 -0.09460515 -0.31237990
#> 49  049  0.36460420 -0.43074855
#> 50  050  1.30069012 -0.06916377
#> 51  051  0.61195586 -1.21244232
#> 52  052  0.71360899  0.13802805
#> 53  053  0.24555734  1.40452887
#> 54  054 -0.44464306 -1.32735169
#> 55  055  1.07197688 -0.27412911
#> 56  056  0.03457314 -0.17074795
#> 57  057 -1.92445992 -1.78559153
#> 58  058 -1.08961601  0.56327038
#> 59  059  0.80644772  0.16139627
#> 60  060 -1.09836122 -1.82070999
#> 61  061 -1.09957169 -0.82281447
#> 62  062 -0.70688192 -0.28444084
#> 63  063 -1.57208447 -1.31324217
#> 64  064  1.25757125  1.52225826
#> 65  065  1.35423546  1.79704251
#> 66  066 -0.28387187  0.16521170
#> 67  067  0.44283057  1.17046012
#> 68  068 -1.58631775 -0.60969008
#> 69  069 -0.74740633  0.10703616
#> 70  070  0.09630203 -1.41778604
#> 71  071  1.37973192  1.65176541
#> 72  072  0.47613917  1.26265927
#> 73  073  0.37719568  0.14667507
#> 74  074  0.39556957  2.01890425
#> 75  075  1.89529784  0.43396582
#> 76  076 -0.37762826  0.97641809
#> 77  077  0.33627991  1.10821315
#> 78  078 -0.74430760 -0.49035454
#> 79  079 -1.31236666 -0.47109398
#> 80  080  1.76916569 -0.29329023
#> 81  081  0.05450092 -1.31924167
#> 82  082  0.89941549 -0.73631822
#> 83  083 -0.14377464 -0.17834902
#> 84  084  0.82438317  0.54014148
#> 85  085  0.78665594 -0.25634597
#> 86  086  0.03761111  1.07866509
#> 87  087 -0.44060016  0.24702346
#> 88  088 -2.49277583 -1.68323976
#> 89  089  1.93395089  1.25529184
#> 90  090 -1.11159094 -0.10641348
#> 91  091  0.42544603 -0.08483844
#> 92  092  1.67854598  0.26095655
#> 93  093 -0.76918822  2.02813323
#> 94  094  0.35935785  0.72306279
#> 95  095  0.64873137 -0.17939464
#> 96  096  0.56610632 -0.61070799
#> 97  097 -0.03880012 -0.86963210
#> 98  098 -0.24911434 -0.28778165
#> 99  099  0.01350811  1.28114010
#> 100 100 -0.46890078 -1.82820586