Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5140236
#> Y_2 0.5140236 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4934350  0.1508986
#> 2  0.5086674 -1.2756442
#> 3 -1.5144136  0.2565382
#> 4 -1.2690572  0.7884830
#> 5 -1.1486811 -3.0274311

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.92574103 -1.26006857
#> 2   002 -1.00609261 -1.62908782
#> 3   003  0.05322739 -0.13405485
#> 4   004  1.54947331 -0.16877923
#> 5   005 -0.78179193  0.69165002
#> 6   006  1.59159627  1.07113022
#> 7   007  0.38710822  0.43634293
#> 8   008  0.10699582  0.27842371
#> 9   009  1.76028445  0.82882474
#> 10  010  0.85442345  1.17205690
#> 11  011 -0.06804687  0.59400978
#> 12  012 -0.39364466 -1.64297407
#> 13  013  0.59609058  0.69698238
#> 14  014  0.68887931  0.88612194
#> 15  015 -0.54572638 -1.23558568
#> 16  016  0.93377912 -0.12404330
#> 17  017  0.20497429 -0.32869471
#> 18  018 -0.75501808 -1.77052801
#> 19  019  0.54924914  1.00878220
#> 20  020  1.01509132 -0.12806627
#> 21  021 -0.36388367 -0.58621184
#> 22  022  0.71755079  0.20126403
#> 23  023 -2.62532394 -1.39737935
#> 24  024  0.49373559 -0.92581198
#> 25  025  0.31427307 -0.28294972
#> 26  026  0.02515504 -0.06905087
#> 27  027  3.75205966  4.10556073
#> 28  028  0.44590116  1.66487729
#> 29  029  0.39158818 -0.35945111
#> 30  030  0.24789191  1.40453344
#> 31  031  1.95904227  0.79396587
#> 32  032 -0.16868828 -2.48517262
#> 33  033  0.28023657  0.55551369
#> 34  034  0.98781047 -0.35953283
#> 35  035 -0.73142200  0.06290750
#> 36  036 -0.24111463 -0.77958520
#> 37  037 -1.20594531 -0.90763748
#> 38  038 -0.52341991 -1.08907705
#> 39  039 -1.13343071 -1.03222821
#> 40  040  0.43841513  1.71014869
#> 41  041 -0.32816451  0.19647964
#> 42  042  0.33673082 -0.76281998
#> 43  043  0.16674433  0.78849589
#> 44  044 -0.64949670  1.41035148
#> 45  045 -1.38154279 -2.18585878
#> 46  046 -0.69501003 -0.11429621
#> 47  047 -0.19541615 -0.18160194
#> 48  048  0.92491356  1.44135634
#> 49  049  0.25664923  2.58478009
#> 50  050 -0.25411507  0.92572153
#> 51  051  0.53731233  0.19595899
#> 52  052  0.78174458  1.59187796
#> 53  053  0.48721931 -1.48673349
#> 54  054  0.97165410 -1.78868305
#> 55  055  2.35957071  2.30660681
#> 56  056 -0.24555330  0.04486860
#> 57  057 -1.13466605 -0.97351586
#> 58  058  1.26400073  1.01724295
#> 59  059 -1.43836065 -1.18668758
#> 60  060 -0.35684316 -0.88517389
#> 61  061  1.18212795  2.53934083
#> 62  062  0.84701523 -0.09776699
#> 63  063 -0.54965064 -0.17111958
#> 64  064  0.35104145  0.44863557
#> 65  065 -1.85648301 -1.78781573
#> 66  066 -1.37690222 -0.39256618
#> 67  067  1.89551877 -0.41854239
#> 68  068  1.11711534 -0.73241005
#> 69  069  0.65430986 -0.77006220
#> 70  070 -0.87735822 -0.31215984
#> 71  071 -0.60116495 -0.97281199
#> 72  072 -0.23220308  0.23094409
#> 73  073 -1.01373959 -0.39518275
#> 74  074  0.55927286  1.08560552
#> 75  075  1.03571939  0.37541948
#> 76  076 -0.82027294 -0.19692079
#> 77  077  1.04823894  0.25180475
#> 78  078 -0.85738752  1.76803611
#> 79  079  0.86251916  1.21286002
#> 80  080  0.30137853  0.27293011
#> 81  081  0.12219533  0.32476883
#> 82  082  0.76473940  0.39521134
#> 83  083 -1.53999184 -0.06429545
#> 84  084 -0.96009437  0.19612995
#> 85  085 -0.96758745  0.03427617
#> 86  086  0.70877608 -0.48834286
#> 87  087 -0.55031788  0.33810501
#> 88  088 -0.16505519  0.34628701
#> 89  089  0.44900230  0.75915335
#> 90  090  1.95904965 -0.34499451
#> 91  091 -0.77434042 -1.10791815
#> 92  092 -1.41006234 -2.01758495
#> 93  093 -1.08178276 -0.08942517
#> 94  094 -0.14395341 -0.31596765
#> 95  095  0.65298464 -0.39723755
#> 96  096  0.96211611  0.71419387
#> 97  097  0.54927671 -1.16344637
#> 98  098 -2.74912602 -0.66625616
#> 99  099 -1.24416415 -0.16358001
#> 100 100  1.52421211  1.18484469

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.61049636  0.114344320
#> 2   002  0.08480902 -0.046756798
#> 3   003  0.90367415  0.285602260
#> 4   004 -2.11718009 -0.344509673
#> 5   005  0.81024762  0.483943924
#> 6   006  0.16130769  0.170496347
#> 7   007 -0.77779013  1.233085344
#> 8   008 -0.33408966  0.071681226
#> 9   009  0.06807363  0.970793018
#> 10  010  1.43430396  1.656603904
#> 11  011 -1.13783328 -1.255812148
#> 12  012 -0.50442080 -1.039096955
#> 13  013 -1.52923654 -2.057019434
#> 14  014 -0.39791372 -1.661777640
#> 15  015 -0.12767585  0.166386575
#> 16  016  0.26662189  0.079480462
#> 17  017 -0.22254445 -0.949255996
#> 18  018 -0.27303191  0.309653771
#> 19  019 -0.72493337 -1.001850916
#> 20  020  0.80029976 -1.114939408
#> 21  021 -0.60491432 -0.088000951
#> 22  022  0.11481889  0.025435092
#> 23  023  0.06923388  0.429994871
#> 24  024  1.48743079  2.330471294
#> 25  025 -1.20367625 -0.560100269
#> 26  026 -0.62283670  0.873451496
#> 27  027  0.27547364  1.749910516
#> 28  028 -0.90609382 -0.567848919
#> 29  029  1.65912838  2.751618166
#> 30  030 -0.59193884 -0.440615788
#> 31  031 -0.69871220  0.144635450
#> 32  032 -0.93071761 -0.147276882
#> 33  033 -0.12833999 -1.917368613
#> 34  034  0.36094155 -0.516171158
#> 35  035  1.45833077 -0.396736130
#> 36  036 -1.34024336  0.312556089
#> 37  037  0.06151336  0.964833915
#> 38  038  1.46665589  2.470122509
#> 39  039 -0.44735670  0.386027003
#> 40  040  1.04982460  0.804682546
#> 41  041  0.16316629 -0.049933226
#> 42  042 -0.90925305 -0.421519140
#> 43  043  1.63423988  0.515731683
#> 44  044 -0.07290458 -0.136071252
#> 45  045  0.01742328  1.244326978
#> 46  046 -0.96642793 -0.381154775
#> 47  047  0.90407783 -1.044993925
#> 48  048 -0.14994982  0.019586467
#> 49  049  0.46525830  0.653217430
#> 50  050 -0.02890494  0.467077840
#> 51  051 -1.11187675  0.074515962
#> 52  052  0.82952402 -0.922137018
#> 53  053 -1.15514170 -0.570307026
#> 54  054  0.39822472 -0.405475036
#> 55  055  0.02898917  1.045752229
#> 56  056  0.22962404 -0.225150844
#> 57  057 -1.36882183  0.107485262
#> 58  058  0.37157629  0.320582294
#> 59  059 -0.58140851 -1.995511179
#> 60  060  1.29099513  1.265877400
#> 61  061 -0.43427139  0.574546229
#> 62  062 -0.10380848 -0.357531280
#> 63  063 -0.19959466 -0.003845818
#> 64  064 -0.59895731 -1.801005732
#> 65  065  0.80341834  1.041643916
#> 66  066 -1.67348056 -0.104910948
#> 67  067  1.77877177  1.164963935
#> 68  068 -1.27540613  0.687079700
#> 69  069  2.54239679  2.576189430
#> 70  070 -1.18252290 -1.665491959
#> 71  071 -0.86507110  0.270469280
#> 72  072  0.55340546  0.017200293
#> 73  073  0.33691029 -0.339572918
#> 74  074 -0.80585681 -0.952720777
#> 75  075  0.54550539 -0.098591641
#> 76  076 -0.25414907 -0.028095005
#> 77  077 -1.25482752 -0.714534691
#> 78  078 -0.56937400 -0.969171040
#> 79  079 -1.43642465  1.145939402
#> 80  080  1.40959084  1.253497521
#> 81  081 -1.32095403 -2.041849133
#> 82  082  0.20303510  0.391610653
#> 83  083 -0.92389060 -0.115523942
#> 84  084  0.11087295 -0.883499166
#> 85  085  0.79724732 -0.418535832
#> 86  086 -0.99790866  0.138931131
#> 87  087 -0.09909516 -0.835074142
#> 88  088  0.04165814  0.078025338
#> 89  089 -0.13465150 -0.167648649
#> 90  090  0.69601943  1.714629112
#> 91  091  1.66965590  1.137604855
#> 92  092  0.22960393  0.173105234
#> 93  093 -1.13685866 -1.245159530
#> 94  094 -1.00271899 -1.557792984
#> 95  095  1.90938797  1.374006549
#> 96  096 -0.19121940 -0.585276653
#> 97  097  1.41489364  1.916449118
#> 98  098 -1.36264518 -0.481894239
#> 99  099 -1.63883572 -1.114753637
#> 100 100  0.82405686 -0.507875514