Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.436606
#> Y_2 0.436606 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4094410  0.1206482
#> 2  0.3935081  0.7260505
#> 3 -0.4289037 -0.5002076
#> 4 -0.6606404  0.9073496
#> 5 -0.5018369  1.6727970

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.493143006 -0.23080943
#> 2   002 -0.339848157  0.83270517
#> 3   003 -0.818846096 -2.08505666
#> 4   004 -0.401450452 -0.13043573
#> 5   005  0.494597059  0.24126153
#> 6   006 -0.977729017 -2.52060056
#> 7   007  0.777451334 -0.09100053
#> 8   008 -0.346517172 -0.86325583
#> 9   009 -0.545403344 -0.21381630
#> 10  010  0.269426022  0.37091365
#> 11  011  0.738621287 -0.75008281
#> 12  012 -0.741271739  1.22447555
#> 13  013 -0.588004938 -0.58531326
#> 14  014  0.472426239 -0.76464078
#> 15  015 -1.024190041  0.63919297
#> 16  016  0.576628190  0.44353999
#> 17  017  0.729963172 -0.23350215
#> 18  018 -0.320770873  1.37098852
#> 19  019  0.968784111  0.46199979
#> 20  020 -0.330162262 -0.49736786
#> 21  021  0.701241128 -0.65885966
#> 22  022  1.157781537  0.50211203
#> 23  023  2.630712963  0.88783007
#> 24  024 -0.752572580 -0.30962805
#> 25  025  0.064225791 -0.54017375
#> 26  026 -0.571395860 -1.33383353
#> 27  027 -1.366929522 -0.50750876
#> 28  028 -0.638528156  0.45305794
#> 29  029 -1.024414712 -0.20193732
#> 30  030 -0.912646367 -1.32069681
#> 31  031  0.982571728 -0.08357984
#> 32  032 -0.513732581 -0.45713056
#> 33  033  0.446084221 -1.96651226
#> 34  034 -0.828762663 -0.59139801
#> 35  035 -0.002357405  1.20919094
#> 36  036  1.733613252 -0.33741100
#> 37  037 -0.978361668  0.52958230
#> 38  038  0.810323233  1.53671140
#> 39  039 -1.773131303 -2.13290037
#> 40  040 -1.129434013 -1.61187999
#> 41  041 -0.397773013 -0.22159917
#> 42  042 -0.435435107 -0.40801894
#> 43  043  0.419033131 -0.91936912
#> 44  044  0.766855401  1.37016494
#> 45  045 -0.657508505 -1.01429564
#> 46  046 -0.716955135 -0.30765657
#> 47  047 -2.214519770 -1.20463966
#> 48  048  2.421098142  0.64307890
#> 49  049  0.264512709  1.20408991
#> 50  050  0.073039966  0.47290882
#> 51  051 -1.827113866 -2.08229120
#> 52  052 -0.577503420 -1.15659148
#> 53  053  0.269173758 -0.01474275
#> 54  054  1.638212784  1.92648199
#> 55  055 -0.208920488 -0.66015814
#> 56  056 -0.173289262  0.32554429
#> 57  057  0.215362210 -0.89709124
#> 58  058 -0.194524598 -0.59875553
#> 59  059 -1.019593515 -1.69707513
#> 60  060 -0.493877509 -0.17374673
#> 61  061  0.630951889  0.82116945
#> 62  062 -0.554615354 -0.66049351
#> 63  063  0.162080642 -0.09359772
#> 64  064 -0.801605618 -1.80449291
#> 65  065 -1.195687320 -1.37305062
#> 66  066  1.122722150  1.79575408
#> 67  067 -0.635977712 -1.90705045
#> 68  068  0.836718739 -1.04276834
#> 69  069  1.600653082  0.31112215
#> 70  070 -0.078075098 -1.44080731
#> 71  071 -0.269521905  0.27348076
#> 72  072  0.009087493  1.16713720
#> 73  073  0.451280614  1.50243778
#> 74  074 -0.039614245 -0.45946739
#> 75  075 -0.585569164  0.78592205
#> 76  076 -0.584893531 -1.22827981
#> 77  077  0.680261749 -0.25717421
#> 78  078 -0.295833772  0.19446028
#> 79  079 -1.082443886  0.13901034
#> 80  080 -1.460736947  0.91310360
#> 81  081  0.239986952 -0.57999323
#> 82  082 -0.871186913 -1.38908377
#> 83  083 -0.126058464 -0.46383198
#> 84  084  1.005045101  2.26265018
#> 85  085 -0.221675619 -0.92023096
#> 86  086  0.732193379 -0.21255100
#> 87  087 -0.160297948  1.77814249
#> 88  088  0.774425611 -0.07386028
#> 89  089  0.497605549  1.29094691
#> 90  090  0.479139849 -2.24148629
#> 91  091  0.601825848 -0.91577578
#> 92  092 -1.622304599 -0.31058127
#> 93  093 -0.642766478 -0.24934912
#> 94  094  0.660735938 -0.56509645
#> 95  095 -0.938409531 -0.48368846
#> 96  096 -0.233299599 -1.81040123
#> 97  097 -0.543079569  0.16805141
#> 98  098  2.006980726  0.40528266
#> 99  099  0.241901830  1.23737839
#> 100 100  1.032643748  0.86560411

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.01485468  0.048113114
#> 2   002  2.05106197  1.178221813
#> 3   003 -1.40800746 -0.154532403
#> 4   004 -1.69047576 -3.323028996
#> 5   005 -1.27296804 -1.999689701
#> 6   006 -0.15301660  0.417093074
#> 7   007 -0.54317247 -0.735197626
#> 8   008 -0.90448102 -0.343865684
#> 9   009 -2.08852589 -1.175749256
#> 10  010  1.27961778  0.312796551
#> 11  011 -1.15230333 -2.009105741
#> 12  012 -0.78342476 -1.115887911
#> 13  013 -1.60408326 -0.559285439
#> 14  014 -1.27828531 -1.003868533
#> 15  015 -1.96670218 -1.470306473
#> 16  016 -0.89438075 -0.324979924
#> 17  017 -0.37549929 -0.912521824
#> 18  018 -0.29652029  0.545623068
#> 19  019  0.02788133 -1.208414117
#> 20  020 -1.61460010 -1.150732063
#> 21  021 -0.55973529 -0.167638258
#> 22  022  0.61039020 -1.016443037
#> 23  023  0.02847084 -1.139519613
#> 24  024 -0.21091451  0.979337324
#> 25  025  0.16583230  0.092040114
#> 26  026  0.46803892  0.228418970
#> 27  027  1.82287328 -0.987883428
#> 28  028  0.84476857 -0.426921743
#> 29  029  1.56496188  0.012009901
#> 30  030  1.89796725  1.994784187
#> 31  031 -0.77391755 -1.166277203
#> 32  032  0.10963027 -0.626735572
#> 33  033  0.90334762 -0.357536419
#> 34  034  0.22028783  0.210838833
#> 35  035 -0.11885604 -0.740797252
#> 36  036  0.57427412  2.562475516
#> 37  037  0.24265021  1.906642137
#> 38  038 -0.68485210 -0.636880104
#> 39  039 -0.59332614 -1.336208655
#> 40  040 -0.82002493 -1.069994735
#> 41  041 -1.25147719 -0.600993741
#> 42  042 -1.41621229 -0.287294883
#> 43  043  1.31462096  0.103169791
#> 44  044 -0.85333679 -1.222178956
#> 45  045 -1.22342540 -0.042903049
#> 46  046  1.09590503  1.356662260
#> 47  047 -1.01482953 -0.835883278
#> 48  048  1.19304596  1.340759642
#> 49  049  0.62395822  1.195449186
#> 50  050  1.04038140  1.448221461
#> 51  051  0.93495170  1.103868214
#> 52  052  0.02237524 -0.147326806
#> 53  053  0.28815039 -0.820150767
#> 54  054  0.92985274  0.691760412
#> 55  055  0.54220313  0.693688275
#> 56  056 -0.12165620  1.980647802
#> 57  057 -0.24588585 -1.622451181
#> 58  058  0.60804173 -1.528807165
#> 59  059 -0.99284101 -0.267131219
#> 60  060  1.05065329 -0.319986571
#> 61  061  1.20554888  1.436856642
#> 62  062 -0.92042711 -0.036277551
#> 63  063 -0.59203502  0.476208505
#> 64  064 -1.22964254 -0.199281223
#> 65  065 -0.49115894 -0.140488167
#> 66  066  0.39662848  0.755159045
#> 67  067  1.24472340  2.136619161
#> 68  068 -0.67533721  0.401591191
#> 69  069  1.53043943  1.658310975
#> 70  070  0.02373527 -0.072401790
#> 71  071 -1.78045850 -1.233055923
#> 72  072 -0.17181051  0.536517923
#> 73  073 -0.14181411 -0.790078312
#> 74  074 -1.02486279 -0.126731393
#> 75  075  0.09669158  0.477133439
#> 76  076 -0.39523958 -1.410691811
#> 77  077 -0.05569441 -0.222351437
#> 78  078 -0.48352058  0.017985037
#> 79  079  1.21798870  0.861394358
#> 80  080  2.49412625  1.720834355
#> 81  081 -0.36290108  0.936808711
#> 82  082  0.42682841 -0.666090067
#> 83  083  0.05227380  0.650058309
#> 84  084  0.35541849  1.110798083
#> 85  085 -0.64180645  1.624006176
#> 86  086 -1.40811797 -0.232298387
#> 87  087  0.89164822  0.502175973
#> 88  088 -1.18440285 -1.012742051
#> 89  089  0.20450531  0.872152701
#> 90  090  0.32944991  0.002652876
#> 91  091  0.13626082 -0.463437348
#> 92  092  1.00685682  0.369644884
#> 93  093  1.06931848 -0.703085077
#> 94  094 -1.00521469  0.679888929
#> 95  095  0.37126116  0.561850505
#> 96  096 -0.16844482 -0.688311452
#> 97  097 -0.65489640 -0.489675494
#> 98  098  0.10444626 -1.629490502
#> 99  099 -0.12383800 -1.041870383
#> 100 100 -0.90219685  0.183427309