Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4924119
#> Y_2 0.4924119 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  0.06357473  1.59002707
#> 2 -0.13133162 -0.03623021
#> 3  1.68140833  1.31704851
#> 4  0.66916449  0.27782497
#> 5  0.73840400  2.01206264

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.90562077  0.33276708
#> 2   002  0.79008647  0.62822332
#> 3   003  1.77964295  1.33061820
#> 4   004 -0.04801559 -0.42673228
#> 5   005  1.54498575  0.75872838
#> 6   006  0.11276019  0.46532257
#> 7   007  0.20012611  0.27923951
#> 8   008 -0.01226470 -0.26717206
#> 9   009 -0.63412549  0.16830305
#> 10  010  0.81356010  0.80835407
#> 11  011  0.66942666 -1.14708548
#> 12  012  1.40035878  1.22420155
#> 13  013  0.10753882  2.26424882
#> 14  014 -0.91653430 -0.99828345
#> 15  015 -1.62473932 -0.01532135
#> 16  016  0.39117445 -0.72154257
#> 17  017 -0.25536055 -0.08151193
#> 18  018  0.43437770  0.33425771
#> 19  019  1.63210925  1.60381627
#> 20  020 -0.30534927 -0.42988653
#> 21  021  0.37671689 -0.37562108
#> 22  022  0.18008114  0.26955722
#> 23  023  0.93528688  0.71757578
#> 24  024  0.58565162  0.14495074
#> 25  025 -0.04166512  0.05665059
#> 26  026  0.35440092 -0.45841751
#> 27  027  0.64495540  1.44547960
#> 28  028 -1.78761569  0.78832551
#> 29  029 -1.52127363 -0.37887686
#> 30  030  0.37704087 -1.53541662
#> 31  031  0.34893120  1.05191343
#> 32  032  1.00131466  0.31973144
#> 33  033  1.31373078  2.05299689
#> 34  034  0.47877179  0.06535382
#> 35  035 -1.29075169 -1.00678133
#> 36  036 -0.25058256 -0.72767148
#> 37  037  0.74586830  0.77915811
#> 38  038  0.57497974 -0.40913205
#> 39  039  0.05095515 -0.34342936
#> 40  040  0.76676931  0.70720117
#> 41  041  0.61791102  1.92939799
#> 42  042  0.13351841 -0.46965570
#> 43  043  0.57924067  1.50677782
#> 44  044  0.33530694 -1.04297554
#> 45  045 -1.06014498 -1.57050769
#> 46  046  1.50874547 -0.76875866
#> 47  047  0.09120437 -0.30440420
#> 48  048  0.70616264  0.68202310
#> 49  049 -0.85326210 -2.67704657
#> 50  050 -0.84042085 -0.34797068
#> 51  051  0.69911462  0.88446691
#> 52  052  0.58965464 -0.27391066
#> 53  053  0.64865027 -0.03377645
#> 54  054  0.49969154  0.52320663
#> 55  055 -1.65723319  0.47014184
#> 56  056  0.95328828  0.82099541
#> 57  057 -0.53495971 -1.49798786
#> 58  058 -0.15792887  1.04558358
#> 59  059 -0.17590542  0.23460928
#> 60  060 -1.00969576 -0.97446174
#> 61  061  0.01805783 -1.12480725
#> 62  062  1.58432565  0.71760004
#> 63  063  0.59199344 -1.01559873
#> 64  064 -1.28373781 -1.12271428
#> 65  065  1.52422333 -0.42637161
#> 66  066 -1.47055484 -0.77418139
#> 67  067  0.25369033  1.54461596
#> 68  068  0.26394898  0.09238940
#> 69  069  1.23475517  1.89157981
#> 70  070  1.22679030  2.84047425
#> 71  071 -1.40187653 -0.82466364
#> 72  072  0.89264642 -1.11677589
#> 73  073 -0.35219036 -1.55731047
#> 74  074 -0.68317315  0.20813682
#> 75  075  1.03841471 -0.62767620
#> 76  076  0.49185669 -0.56445683
#> 77  077  2.05128933  0.02508815
#> 78  078 -1.73780536  0.00121780
#> 79  079 -1.06065506  0.58905046
#> 80  080  0.69842578 -0.70791592
#> 81  081 -0.65446388 -0.65619250
#> 82  082  1.98022699 -0.44572266
#> 83  083  0.69655396  0.32815175
#> 84  084 -0.09069306 -0.73092003
#> 85  085  1.57423231  1.45088025
#> 86  086 -1.96007934  0.77259246
#> 87  087 -0.81014057 -2.84761349
#> 88  088  0.28463975  0.52370638
#> 89  089 -0.59994275 -0.31503732
#> 90  090  1.21789840 -1.04737501
#> 91  091 -0.53289453  0.25551047
#> 92  092 -0.44860822 -0.33421695
#> 93  093 -0.47017397  0.29041997
#> 94  094  0.59306022  1.34174581
#> 95  095 -0.88269486  0.48713392
#> 96  096  0.07903634 -0.22046303
#> 97  097  0.06405991  0.66476811
#> 98  098  0.64379185  0.00666565
#> 99  099 -0.29638755 -0.09745274
#> 100 100 -0.07738987 -0.41195292

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.67283615  0.53553462
#> 2   002 -1.30737786  1.88326307
#> 3   003 -1.35400627 -0.18033628
#> 4   004 -2.08009944  0.33288377
#> 5   005  1.76629410  0.46156241
#> 6   006 -1.71084949  0.22830597
#> 7   007  0.74028304  0.28504092
#> 8   008 -1.07399073  0.41189539
#> 9   009 -0.67977610 -0.48619968
#> 10  010 -0.22793585  0.88488040
#> 11  011 -0.38916048  0.32670564
#> 12  012  0.90490249  0.38473431
#> 13  013  0.49183402  0.72591510
#> 14  014  0.49123022  0.40901221
#> 15  015 -1.33496487 -0.60245819
#> 16  016 -0.12618537  0.50698464
#> 17  017  0.14404984  0.52632729
#> 18  018  0.24848876  0.85920846
#> 19  019 -0.25701264 -0.77082288
#> 20  020 -1.98790219 -1.40306066
#> 21  021 -1.35897789 -1.19513720
#> 22  022 -0.57180389 -0.22539645
#> 23  023 -0.24761647 -0.23842020
#> 24  024 -0.07885495  1.86876428
#> 25  025  0.36343474 -0.25175358
#> 26  026  0.02755110 -0.72754652
#> 27  027 -0.17469879  0.43058938
#> 28  028 -0.58941929  1.25979889
#> 29  029  2.70753191  2.20105803
#> 30  030 -0.84806544  0.46300867
#> 31  031 -1.28436518 -1.10656978
#> 32  032  0.75640239 -0.67809312
#> 33  033 -0.37006799 -0.17413954
#> 34  034  0.38641781  0.14615764
#> 35  035  1.22875119 -0.13811871
#> 36  036  0.99599691  0.52763356
#> 37  037  1.16839935  1.72456613
#> 38  038  0.51524017  0.12867008
#> 39  039  0.19094309  0.54376765
#> 40  040  0.62870622 -0.28470144
#> 41  041 -0.77363798 -1.87279408
#> 42  042  0.18474192  0.25265118
#> 43  043 -1.69222368 -0.27788551
#> 44  044 -0.78416619 -1.78387746
#> 45  045  1.10433127  1.41318792
#> 46  046 -0.16404897  1.07240572
#> 47  047 -1.20034847 -0.39119691
#> 48  048 -2.83882496 -1.59415495
#> 49  049  0.39101380  0.85879042
#> 50  050 -0.90671662  0.57486750
#> 51  051  2.47899313  2.04603474
#> 52  052  0.59449259 -1.00538174
#> 53  053 -0.43379485  0.04675823
#> 54  054 -0.50571831  0.28924341
#> 55  055  0.54853639  1.14717481
#> 56  056  0.10463268 -0.33276907
#> 57  057 -0.22694142  1.72986449
#> 58  058 -0.08082504  0.36419607
#> 59  059  0.87007084 -1.19049399
#> 60  060  0.34128360 -1.00990245
#> 61  061  0.19260934 -0.10818081
#> 62  062 -0.71898300 -0.76082270
#> 63  063  1.19957171 -0.47150355
#> 64  064 -1.18851108 -0.02522940
#> 65  065 -1.56289886  0.42834477
#> 66  066  0.02040348 -3.11303775
#> 67  067  0.60087210  2.30188092
#> 68  068 -0.58630616  0.04522693
#> 69  069  0.45214419  0.71886637
#> 70  070 -1.06760346 -1.27255135
#> 71  071 -1.67875593  0.04910389
#> 72  072 -0.67899566 -0.74169561
#> 73  073  0.22360873 -0.45986709
#> 74  074 -0.61951209 -0.50980683
#> 75  075 -0.72402562 -0.88067019
#> 76  076  0.60291066  1.06163843
#> 77  077  2.03157313  0.50111341
#> 78  078 -1.18334823 -0.13636031
#> 79  079  0.09359197  1.58271346
#> 80  080  0.04556311  0.04426305
#> 81  081  0.45880575 -1.49503036
#> 82  082 -1.44052160 -1.09889545
#> 83  083 -0.39286465 -0.59217578
#> 84  084 -2.84398183 -2.12817430
#> 85  085  2.02920814  1.46596703
#> 86  086 -0.17110822  0.29855925
#> 87  087  1.59895289  0.02879689
#> 88  088 -0.49605307  1.03182139
#> 89  089 -0.88159565 -1.20591444
#> 90  090  1.13647368  2.25769784
#> 91  091 -0.48997656  1.27791500
#> 92  092  1.99234137 -0.34082345
#> 93  093 -1.18562358  1.48301210
#> 94  094 -0.33692442 -1.74487739
#> 95  095 -2.58543571 -1.39470174
#> 96  096 -0.44417541 -0.58507848
#> 97  097 -0.55455717 -1.39071739
#> 98  098  2.04382123  0.73850660
#> 99  099 -1.32433160 -0.69731798
#> 100 100  0.59680876  2.24232935