Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5350203
#> Y_2 0.5350203 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.6421489 -0.09861274
#> 2  0.4120013  0.97207059
#> 3  0.2097289  1.72524909
#> 4  2.3961669  0.84479127
#> 5 -1.5052623 -0.80750086

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.980172560  2.278193083
#> 2   002  1.936681867  0.619857697
#> 3   003  0.874603189  0.449338737
#> 4   004 -0.774714663  0.112549066
#> 5   005  0.517697958  0.412935973
#> 6   006  0.336714954  0.718426075
#> 7   007 -0.381579274 -0.002599278
#> 8   008  0.578735760 -0.351892920
#> 9   009  0.649844802  1.565601946
#> 10  010 -0.769444267  1.462063533
#> 11  011  1.484360332 -0.157171797
#> 12  012 -0.172774271 -0.574193368
#> 13  013 -0.322989065 -0.710051382
#> 14  014 -0.420973362 -0.061369463
#> 15  015  0.013449234 -0.243045858
#> 16  016  0.361377304  0.064237664
#> 17  017 -0.526948688 -0.429365519
#> 18  018  0.090041235  0.891932262
#> 19  019  1.684982387  1.036241834
#> 20  020  0.947450339  1.892269293
#> 21  021 -1.897473200 -1.500308863
#> 22  022  0.236135332  0.445815364
#> 23  023  0.792201551 -0.599510994
#> 24  024  0.016071758 -0.625967005
#> 25  025 -0.002830903  1.085472658
#> 26  026 -0.815157749  0.832525736
#> 27  027 -1.019690601  0.724753664
#> 28  028  0.921795400 -0.100854013
#> 29  029  1.241058899  0.226034889
#> 30  030  0.747937725  1.639332247
#> 31  031  0.473959694 -0.208867263
#> 32  032  0.877641837  0.156376443
#> 33  033 -0.198312986 -0.487161384
#> 34  034 -0.195140695  0.728770721
#> 35  035 -1.050145775 -0.620676263
#> 36  036 -0.454338850 -0.906095101
#> 37  037 -0.757406881 -0.316654522
#> 38  038 -1.783885293  0.945683680
#> 39  039  1.249452135  1.254905273
#> 40  040 -0.106584354  1.228703326
#> 41  041 -0.789482762  0.205241389
#> 42  042 -0.806450200 -0.818386141
#> 43  043  0.489090795 -0.135284140
#> 44  044  0.882402363 -0.947591225
#> 45  045  0.045814361  0.235416959
#> 46  046  1.178507606 -0.335197471
#> 47  047 -0.225903031 -0.819834966
#> 48  048  0.443566215  0.784837931
#> 49  049  1.956809295  1.834599345
#> 50  050  1.556376622  0.326544412
#> 51  051 -0.851620910 -0.352792017
#> 52  052 -2.944077256 -2.882977404
#> 53  053 -0.367212684  0.638967870
#> 54  054  0.319356925  0.137857433
#> 55  055  0.888332828  0.815225041
#> 56  056 -1.033714337 -0.463293909
#> 57  057  0.191197435  1.072657122
#> 58  058  1.118973063  0.071078369
#> 59  059  0.127243195  0.406734302
#> 60  060 -0.829152571 -0.971303699
#> 61  061 -0.358484502 -1.010163705
#> 62  062 -0.457468423  0.875783516
#> 63  063  1.239320293  1.086685364
#> 64  064  0.064469429 -0.687058785
#> 65  065 -0.719130355 -0.075481132
#> 66  066 -1.641859529 -0.164026446
#> 67  067 -1.041776259  1.372721096
#> 68  068 -1.238475724 -0.006142876
#> 69  069  1.022896951  0.172836232
#> 70  070  0.141764221  0.477454273
#> 71  071 -0.181446000 -1.357520952
#> 72  072 -1.478888831 -3.051897698
#> 73  073 -0.087618792 -0.178243122
#> 74  074  0.622259460 -0.962344672
#> 75  075  0.560473128  0.980965085
#> 76  076  0.048371642  0.484639581
#> 77  077 -0.080830631  1.038210484
#> 78  078  0.641644287  1.772214382
#> 79  079  0.084346352 -0.792801847
#> 80  080  1.403963278  0.601980090
#> 81  081  1.592766169 -0.347866435
#> 82  082 -0.911786511 -0.069181110
#> 83  083 -0.814471660 -0.702600963
#> 84  084  1.081509382  1.673462107
#> 85  085 -0.042511004  0.696104573
#> 86  086  0.951039180  1.526253792
#> 87  087 -0.587019545 -0.535873965
#> 88  088 -0.118958925 -0.146961559
#> 89  089  0.375708837  0.546835742
#> 90  090 -0.223263020  0.067362163
#> 91  091  0.452033839 -0.149759750
#> 92  092  2.241188360  1.240307854
#> 93  093 -1.067230600 -1.600518481
#> 94  094 -0.157804620  0.047114349
#> 95  095  1.130592301  1.913224319
#> 96  096 -0.364236262  0.471402261
#> 97  097 -1.431990562 -0.986754401
#> 98  098 -1.097161008 -1.351579767
#> 99  099 -0.825952717 -0.962520506
#> 100 100 -0.266137523  0.301894155

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.908123950 -0.111374290
#> 2   002 -0.802861509 -0.875360576
#> 3   003  0.795001422  0.321162714
#> 4   004 -0.298266591  0.459826573
#> 5   005  1.806700665  1.993872727
#> 6   006  0.871892873  0.501675980
#> 7   007 -0.202857932  1.458593355
#> 8   008  1.115260166  0.507172665
#> 9   009  0.806988273 -0.257022737
#> 10  010  0.313370432 -0.170797950
#> 11  011  0.963843357  0.421431511
#> 12  012  1.441810571  0.037023509
#> 13  013  2.163420419  2.465934882
#> 14  014  0.712380414 -0.371227458
#> 15  015  0.429834095 -0.220900228
#> 16  016 -0.703776441 -0.765586541
#> 17  017 -0.043946980  0.794053322
#> 18  018  0.101446909  0.854591031
#> 19  019 -0.744467657  0.418325490
#> 20  020  1.054916613  0.841063847
#> 21  021  0.865338896 -0.094545477
#> 22  022 -0.857592690  0.553906241
#> 23  023  1.602539245 -0.064460534
#> 24  024 -1.761673538 -2.553066968
#> 25  025  1.309291527  0.710022434
#> 26  026  0.507117375 -0.850703767
#> 27  027  0.840709033  1.501338694
#> 28  028  1.342165746  1.262214240
#> 29  029  0.853957069 -0.454084668
#> 30  030 -0.218256798 -2.038567358
#> 31  031  0.369672651 -0.152925481
#> 32  032  1.212537507  0.223952456
#> 33  033 -0.137342126  1.566532835
#> 34  034  0.037099398  0.830933338
#> 35  035  1.610527190  1.226493653
#> 36  036  0.930020619 -0.156563611
#> 37  037 -0.129087631 -0.068892945
#> 38  038 -0.873999451  0.376992861
#> 39  039 -0.288005454  1.454197907
#> 40  040 -0.309399117 -0.838789799
#> 41  041  0.544423044  1.676819382
#> 42  042 -0.821627396  0.604537875
#> 43  043 -1.574368062 -1.416997499
#> 44  044 -0.834805301 -0.692061811
#> 45  045 -2.400656790 -0.769605295
#> 46  046  0.128611912 -0.008900634
#> 47  047 -0.815520410 -1.144356456
#> 48  048  1.096474464  0.446239352
#> 49  049 -0.903952987 -2.382656232
#> 50  050  1.441410018  1.239549988
#> 51  051  0.024578709  0.409561531
#> 52  052  0.334401928  0.374584397
#> 53  053 -0.383200034  0.076832545
#> 54  054  1.341354181 -0.763032880
#> 55  055 -0.399987088  0.398519305
#> 56  056  3.221948284  2.407998152
#> 57  057  0.890647685  1.180112498
#> 58  058  0.101026564  0.592847628
#> 59  059  0.216839167  1.882073270
#> 60  060 -2.535423022 -1.299646595
#> 61  061 -0.297371856  0.194925251
#> 62  062 -0.839368631 -1.220788697
#> 63  063 -0.075873989 -0.385133208
#> 64  064 -0.656437513 -0.089580834
#> 65  065 -2.330283353 -1.298460600
#> 66  066  0.555891785 -0.882171465
#> 67  067 -0.083762730 -0.349054408
#> 68  068 -0.590915361 -0.880639350
#> 69  069 -0.072431918 -2.153008133
#> 70  070  0.528887740  2.193659067
#> 71  071 -1.637181637 -0.868789926
#> 72  072 -0.761203880 -0.427618205
#> 73  073  0.410087582 -0.147342878
#> 74  074 -2.299866579 -0.157991013
#> 75  075 -1.045282683  0.756864761
#> 76  076  0.367132607  0.337408683
#> 77  077  0.294702829 -0.701334678
#> 78  078  0.601758649  0.883038087
#> 79  079 -1.136552385 -0.956021569
#> 80  080 -0.867936569 -0.152200504
#> 81  081 -1.497490054 -1.453265995
#> 82  082 -0.990387896 -1.250008517
#> 83  083  0.397695005  1.475194205
#> 84  084  0.999686836  0.232964598
#> 85  085 -0.154040452 -0.498090188
#> 86  086  2.171881290  0.222336615
#> 87  087 -0.880212332  0.263659531
#> 88  088 -0.859942252 -0.312303663
#> 89  089 -0.752142626 -0.079738329
#> 90  090  2.050785917  0.838665662
#> 91  091  0.080710155 -1.021402642
#> 92  092  1.188926142  0.169253433
#> 93  093 -1.273393643 -0.335727663
#> 94  094  1.078072146 -1.340473288
#> 95  095  0.539831069  0.795150952
#> 96  096 -0.028627959 -1.042323408
#> 97  097  1.485953277  0.712986389
#> 98  098  1.467866221  0.804202575
#> 99  099 -0.627360952 -0.193096706
#> 100 100  0.006962068 -1.191235598