Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5002933
#> Y_2 0.5002933 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5348906 -0.6266458
#> 2  1.0819571  0.1951916
#> 3 -0.1120145  0.6439935
#> 4 -0.4240264  2.1834560
#> 5  1.1045973  1.6992168

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.773944795  1.840004877
#> 2   002  1.296609066  0.714356923
#> 3   003  0.281409075  0.979999243
#> 4   004  0.319604483  0.191889120
#> 5   005  0.040994040  0.497192118
#> 6   006  1.935603592  1.378251119
#> 7   007 -0.058895242  0.058325002
#> 8   008  2.000334262 -0.846457957
#> 9   009 -0.703788383  0.649977132
#> 10  010  0.316468719  0.795320487
#> 11  011 -0.639176478  0.637556711
#> 12  012  0.175863474 -0.535388511
#> 13  013 -0.518127604  0.717237540
#> 14  014 -0.008370616 -0.615608375
#> 15  015  0.213431414 -0.102628575
#> 16  016  0.368506435  0.255919967
#> 17  017 -0.802148871  0.108371721
#> 18  018 -0.267375565  0.089577905
#> 19  019  0.467256798 -2.044530353
#> 20  020 -0.316623593  0.337462149
#> 21  021  0.661022561  1.568153666
#> 22  022  1.836976013  1.095259845
#> 23  023  0.423226154  0.335672402
#> 24  024 -0.208554000  0.940586266
#> 25  025  0.978230875  0.981327461
#> 26  026  1.006871069  0.928281610
#> 27  027 -1.035011638 -1.482110289
#> 28  028 -0.369379041 -0.877195201
#> 29  029  0.951776055  1.999319733
#> 30  030 -1.629629797 -1.196764422
#> 31  031 -0.555435271 -0.266167247
#> 32  032 -2.324555864 -1.144842281
#> 33  033 -0.090683343  0.620007347
#> 34  034  1.508811710  0.579764257
#> 35  035  0.201889870  0.908905046
#> 36  036 -0.266911771 -0.605086518
#> 37  037 -0.097906156  0.047220411
#> 38  038 -0.004023925 -2.185948067
#> 39  039 -1.138055060 -0.087761426
#> 40  040  0.432953023 -2.630699896
#> 41  041  0.737707371 -0.462092697
#> 42  042 -0.690210705 -1.011283040
#> 43  043  0.917264510 -0.073444470
#> 44  044 -0.929358170 -2.679404059
#> 45  045  0.392299002  1.212594036
#> 46  046  2.253736183  2.422542071
#> 47  047 -0.113308913  0.474423525
#> 48  048  0.437307813  0.408209597
#> 49  049  1.084533372 -0.140699423
#> 50  050 -2.751652514 -1.171597683
#> 51  051 -0.122264680 -1.108881701
#> 52  052  0.142482188 -0.680803818
#> 53  053 -0.270599330 -0.196765247
#> 54  054 -0.932747656 -0.999315696
#> 55  055  0.055373064  0.031759514
#> 56  056 -0.469549387  0.228848287
#> 57  057 -1.006203171  0.251614768
#> 58  058  0.887745538  0.806083996
#> 59  059  2.330298016  1.610152270
#> 60  060 -1.592644582 -1.400788298
#> 61  061  0.794555083 -0.289374359
#> 62  062 -0.237233145 -0.585567095
#> 63  063  0.585669389  0.619869736
#> 64  064 -0.822867809 -0.069420324
#> 65  065  1.639623940  1.486333801
#> 66  066 -0.182471344  0.256769324
#> 67  067  0.037368500 -1.188037799
#> 68  068  0.749275719  1.603178089
#> 69  069  0.817263831  0.709964047
#> 70  070  0.584902270  1.170414874
#> 71  071 -1.230729795 -1.613225230
#> 72  072  1.245580431  2.928501028
#> 73  073  0.573944224 -0.612773472
#> 74  074 -1.242565342  0.310963224
#> 75  075 -1.917000113 -0.546992613
#> 76  076  0.038839017 -0.539259861
#> 77  077 -0.467121301  0.151676497
#> 78  078 -0.107154870  0.200105471
#> 79  079 -0.731610268 -0.006065927
#> 80  080 -0.241243191 -0.910223817
#> 81  081  0.650723336  1.779221825
#> 82  082  0.876803590  1.244879627
#> 83  083 -2.009051316 -0.942314251
#> 84  084 -0.312502265 -0.074603942
#> 85  085  2.212274604  1.090381773
#> 86  086 -1.072764784 -0.051342090
#> 87  087  0.770560767  0.823769180
#> 88  088  0.008166741 -0.430858740
#> 89  089 -0.324256167 -0.774644205
#> 90  090  0.527160168 -1.030141043
#> 91  091  1.921317113  0.404881050
#> 92  092  0.822094199  2.277427980
#> 93  093 -0.897220209  0.830697212
#> 94  094 -1.084024589 -1.040484687
#> 95  095 -1.091114928 -0.773142688
#> 96  096 -0.493871223 -0.366357579
#> 97  097  0.393427978  0.511304835
#> 98  098 -1.133895857  0.284320352
#> 99  099  0.039413133 -0.860549336
#> 100 100  0.853601154  1.438325727

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -2.134014910 -1.071894201
#> 2   002  1.914805569  1.082589920
#> 3   003  0.458148219 -0.797449237
#> 4   004 -0.378912874 -1.277465149
#> 5   005  1.820679803  1.094529970
#> 6   006 -0.638100829 -0.134677214
#> 7   007  0.351609016 -0.969745776
#> 8   008  0.721655004 -0.023759737
#> 9   009  2.144172910  1.330030282
#> 10  010 -2.069267795 -1.826004216
#> 11  011  0.100537496 -1.214410758
#> 12  012  0.926332419 -0.163027000
#> 13  013 -0.994316372  0.026274024
#> 14  014  0.626178262 -0.109893056
#> 15  015 -1.241148926 -1.017905109
#> 16  016  1.258295613  0.328711762
#> 17  017 -0.730958779 -0.290323644
#> 18  018 -0.050777504  0.259530963
#> 19  019  1.135844085  0.754060215
#> 20  020  0.219982211 -0.340988098
#> 21  021  0.085821106  0.976457869
#> 22  022  0.044834489 -0.946414849
#> 23  023 -0.124859890 -1.123652582
#> 24  024  0.727590714 -0.333709276
#> 25  025  0.005767333  1.737740479
#> 26  026  1.277065186  0.309758592
#> 27  027  1.635537845 -0.243053460
#> 28  028  1.064871861 -0.172236465
#> 29  029 -0.888639283 -1.028512293
#> 30  030 -1.472589288 -1.357478777
#> 31  031  0.818113783  0.185576895
#> 32  032 -0.363312157  0.093790601
#> 33  033  1.497466310 -0.056278248
#> 34  034  0.610090523  0.386310325
#> 35  035  3.235180583  1.963321976
#> 36  036  0.214696337  0.225720287
#> 37  037  0.178785418 -1.622424897
#> 38  038  0.418366071  0.860040570
#> 39  039 -2.600118907 -1.753337738
#> 40  040 -0.228047543  0.747906107
#> 41  041  0.839091564  0.034563747
#> 42  042  0.576750878  0.483756435
#> 43  043 -0.347424674 -0.003711707
#> 44  044 -1.143955524 -1.197833668
#> 45  045 -0.991129768 -0.805472197
#> 46  046 -0.204592634 -0.195802498
#> 47  047 -0.565411122 -0.375487766
#> 48  048 -1.861035909 -0.693862939
#> 49  049  1.086835849  0.374442760
#> 50  050 -0.994299746 -0.391457763
#> 51  051 -0.490890125  1.804323437
#> 52  052  0.504852029  0.906397987
#> 53  053  0.034725632  0.131691352
#> 54  054 -2.936540292 -1.394180870
#> 55  055  1.794388001  0.441909281
#> 56  056 -0.910069758 -1.423226041
#> 57  057  1.513416008  1.638055260
#> 58  058 -0.327543779 -0.527144903
#> 59  059 -0.890352091 -0.709440256
#> 60  060 -1.639576653 -2.054630643
#> 61  061 -0.896033366 -0.072644429
#> 62  062 -1.025464313 -1.284727095
#> 63  063  1.274057141  0.484935032
#> 64  064  0.188655839 -0.939018871
#> 65  065 -0.039995720  0.240301006
#> 66  066 -0.169194771 -1.731979652
#> 67  067  0.359133448 -0.447756461
#> 68  068 -0.153694597 -0.023402823
#> 69  069  1.185689719  0.809607459
#> 70  070 -0.761016656  0.143418194
#> 71  071  0.899485289  0.260565996
#> 72  072  1.394021333  1.327161979
#> 73  073  0.537568791 -0.340398637
#> 74  074 -1.305950058 -1.346863639
#> 75  075  1.732297929  1.102658039
#> 76  076  1.610144122  0.065543503
#> 77  077 -0.622501651 -0.497924546
#> 78  078  0.442200644  1.284828647
#> 79  079  0.995788042 -1.010530700
#> 80  080 -1.010181359 -1.077530828
#> 81  081 -0.639419766  0.382760556
#> 82  082 -1.841044351  1.314377920
#> 83  083  0.180818796 -0.277223517
#> 84  084  1.503323650  1.045969790
#> 85  085 -0.812702288  1.063019861
#> 86  086  0.546732956  0.344155484
#> 87  087  1.318131134  2.366266451
#> 88  088 -0.162306616 -0.097727325
#> 89  089 -0.230623962  0.891546354
#> 90  090 -1.044403834  0.013079342
#> 91  091 -1.275268429 -1.436479440
#> 92  092  0.289045512 -0.657497676
#> 93  093 -0.782212726 -0.766898704
#> 94  094  1.187084512  2.043782344
#> 95  095 -0.325050186 -0.160648605
#> 96  096 -1.478863917 -2.883414689
#> 97  097 -0.842683860 -1.364486040
#> 98  098 -1.274387162  0.765638211
#> 99  099 -0.679621233 -1.259731172
#> 100 100  1.144717494  0.377954941