Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4672472
#> Y_2 0.4672472 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.8318949 -0.3786993
#> 2  0.1750820  1.1068214
#> 3  0.5176826  0.4997361
#> 4 -1.5345247 -0.9001168
#> 5  0.0968348  1.7405006

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.01685685 -0.809792682
#> 2   002 -0.42038491 -0.534365788
#> 3   003  0.18724216  0.380276273
#> 4   004  0.50992488  1.049359631
#> 5   005 -0.26180180  0.241813477
#> 6   006 -0.52852971  1.186204132
#> 7   007 -0.09990527 -1.569055985
#> 8   008  0.10112308  1.065497415
#> 9   009 -2.36942314 -2.175298493
#> 10  010  0.20770948 -1.009956518
#> 11  011 -0.90481898 -1.499442819
#> 12  012 -0.75658627  2.044432794
#> 13  013  0.69355573  0.487274795
#> 14  014 -1.74271926 -1.245172100
#> 15  015 -0.99642641 -1.680918778
#> 16  016  1.22153938  2.348189034
#> 17  017 -1.30612706  0.612463932
#> 18  018  0.44780350 -0.443813431
#> 19  019  1.78684286  1.378836935
#> 20  020 -0.28988733 -0.256566403
#> 21  021 -0.11203758 -0.514611967
#> 22  022  1.24103359  1.265931826
#> 23  023 -0.23331390  0.066478738
#> 24  024 -1.12715877 -1.191986185
#> 25  025 -0.39948652 -0.828281721
#> 26  026  0.22729536  0.307625869
#> 27  027  1.03941777  1.296139092
#> 28  028 -0.55153701 -0.813457942
#> 29  029  0.42607249  0.603995144
#> 30  030  0.77550380 -0.024659895
#> 31  031 -0.76100512 -0.418183499
#> 32  032 -0.73082837 -1.724603892
#> 33  033  0.98564563 -0.664961546
#> 34  034  0.04627289  1.949990204
#> 35  035 -1.22286003 -0.646294089
#> 36  036  0.54633686 -0.115035663
#> 37  037 -1.47459629 -0.246974647
#> 38  038  1.07307308  2.183749915
#> 39  039 -0.02597124 -0.275180190
#> 40  040 -0.33116775 -0.218981793
#> 41  041  0.27285177  0.775723532
#> 42  042  0.74555901  0.557737497
#> 43  043 -1.56397092 -2.102805013
#> 44  044 -0.28675326  2.108543309
#> 45  045  1.27524041  0.433489711
#> 46  046 -1.06851606 -1.593113257
#> 47  047  2.26552144  0.257147510
#> 48  048 -1.66097637 -0.303473783
#> 49  049  0.11882273 -0.476782516
#> 50  050  2.27513813  0.011745822
#> 51  051  0.16275691  0.362892717
#> 52  052  0.14375767  1.266762481
#> 53  053 -0.04087211 -1.075598174
#> 54  054 -0.28628262 -0.107101080
#> 55  055  1.13629114  0.871462972
#> 56  056  1.33984249  0.662821908
#> 57  057  0.54599936  0.135559781
#> 58  058  1.45072397  0.358369321
#> 59  059 -1.02084639 -1.617230583
#> 60  060 -0.04629519 -1.212038101
#> 61  061 -0.55513038  0.168844759
#> 62  062 -0.15678674 -0.918310925
#> 63  063 -0.02096244  0.572957635
#> 64  064  1.69131475  1.078737424
#> 65  065 -0.85596759 -1.760187366
#> 66  066  0.28159315 -0.283991818
#> 67  067 -1.91639521 -0.272707820
#> 68  068  1.84709749  0.888076020
#> 69  069  0.96350310 -0.654310216
#> 70  070 -0.05293929  0.605019447
#> 71  071  0.38746421  0.334256215
#> 72  072 -0.88037733  0.166818812
#> 73  073  1.19061226 -0.731426313
#> 74  074 -0.35287066  0.184591939
#> 75  075  1.65803205  1.680662844
#> 76  076 -1.76480260 -1.365138197
#> 77  077  0.71575454  0.027880522
#> 78  078 -0.10230433  0.945221962
#> 79  079  1.11643706 -0.048853138
#> 80  080  0.36663482  0.529486222
#> 81  081 -0.66697088 -1.401232547
#> 82  082 -0.44453714  1.121444590
#> 83  083  0.53646962 -0.049884409
#> 84  084 -0.42362868 -1.936687267
#> 85  085  0.52778076  0.303666339
#> 86  086 -0.99679050  0.498191878
#> 87  087 -0.92853167  0.007238305
#> 88  088  0.37378135 -0.012299778
#> 89  089  0.33011614  0.750986901
#> 90  090  0.56519265  0.552631949
#> 91  091 -0.34622791  0.280820933
#> 92  092 -2.02342317 -0.736578087
#> 93  093 -0.87833658 -1.399504298
#> 94  094 -1.11231951 -0.221564165
#> 95  095  0.37155063  0.125339108
#> 96  096 -1.05797706 -2.587584043
#> 97  097  0.04029782  0.349106424
#> 98  098  1.68546174  1.703127466
#> 99  099  0.53319539  0.895007819
#> 100 100  3.15440401  1.254068792

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.48337995  1.366867170
#> 2   002  1.09944108  0.749843616
#> 3   003 -1.63422215 -0.358127989
#> 4   004  0.19313871 -0.266994039
#> 5   005 -0.43789031 -0.748328581
#> 6   006  0.18038608 -0.799537028
#> 7   007  1.69770661  0.526200269
#> 8   008  0.63666278  0.766652481
#> 9   009 -0.25385077  0.369642575
#> 10  010  0.44288804 -0.000752075
#> 11  011 -0.70417715 -0.586992433
#> 12  012 -0.49948176  0.446700559
#> 13  013 -0.80581295 -0.704821915
#> 14  014  0.04129675 -1.143664024
#> 15  015  0.74147054 -0.051801361
#> 16  016 -1.77886544 -1.864472269
#> 17  017 -1.61218693 -0.150038105
#> 18  018  0.37336896 -0.456306878
#> 19  019  0.30701840  0.332137733
#> 20  020 -0.02232813  1.055361357
#> 21  021 -0.06943737  1.454149010
#> 22  022 -0.52316558 -0.429472968
#> 23  023  1.07151875  1.963154015
#> 24  024 -0.58672085  1.351147203
#> 25  025 -0.31220147  0.673037818
#> 26  026  0.50591446 -1.163741075
#> 27  027 -0.89740079  0.271987100
#> 28  028  0.61379865  0.140861612
#> 29  029  0.21377941  0.655971547
#> 30  030 -1.50415726 -0.672536257
#> 31  031  0.72057259 -0.783117664
#> 32  032 -1.24412497  0.152583131
#> 33  033  0.64022762  0.366158782
#> 34  034  0.68552524 -0.562927674
#> 35  035 -1.63136708 -1.317120980
#> 36  036 -0.42140685 -0.435047694
#> 37  037 -1.10819707 -0.201758289
#> 38  038 -2.07568880 -1.300846438
#> 39  039  0.50981954  1.155716122
#> 40  040  0.12359492 -0.522020892
#> 41  041  0.34976826  0.697918357
#> 42  042 -1.26694851 -0.526904516
#> 43  043 -1.59433552 -1.193116260
#> 44  044  0.53711343  0.425699181
#> 45  045  1.24710910  2.222184383
#> 46  046 -0.90304829 -0.119390413
#> 47  047  0.56217918  0.328104238
#> 48  048  0.57270333 -1.043969521
#> 49  049 -1.37235073 -0.250412665
#> 50  050  0.88573729 -0.560346233
#> 51  051 -0.64746784 -1.159356218
#> 52  052 -0.22525421 -1.883180601
#> 53  053 -2.13824486 -1.144936001
#> 54  054  1.67206516  1.798078050
#> 55  055  0.36863149  0.328472643
#> 56  056 -0.44938079  1.073949372
#> 57  057  0.47503229  1.557020863
#> 58  058  0.08584212 -0.271278157
#> 59  059  0.67591256 -0.834245335
#> 60  060  0.30638174  1.146399562
#> 61  061  0.63971738  0.140794941
#> 62  062 -1.55726662 -0.718969302
#> 63  063 -0.20174554  1.074297728
#> 64  064  2.26874167  1.668021580
#> 65  065 -0.27063967 -1.876888165
#> 66  066 -0.69566625 -0.045511599
#> 67  067  0.58549715  0.701768548
#> 68  068 -0.39648207  0.110163242
#> 69  069  0.04417942  2.150845622
#> 70  070 -0.69137112 -1.470648754
#> 71  071 -0.61809111  0.998548612
#> 72  072  0.18314836  1.345009762
#> 73  073  0.36119228  0.038675969
#> 74  074 -0.06386743 -0.994653932
#> 75  075 -0.89099736 -0.503544960
#> 76  076  0.40462465 -0.466653888
#> 77  077  0.12264456 -0.180290694
#> 78  078 -1.81400614 -1.415259708
#> 79  079  1.08024607  0.191345418
#> 80  080  0.59364677 -0.264275633
#> 81  081  0.26423675 -0.106690095
#> 82  082  0.43329481  1.415463491
#> 83  083  0.38961045  0.194144532
#> 84  084  0.29562463  0.354161322
#> 85  085  1.28089558  2.119764399
#> 86  086 -0.26322498 -0.516676355
#> 87  087 -0.83986999 -1.244327127
#> 88  088  0.21237533  0.851426810
#> 89  089 -0.06166609 -0.622390114
#> 90  090  0.40936661  1.030609676
#> 91  091 -0.86210175 -1.558945702
#> 92  092  2.05159080  0.605002863
#> 93  093  0.51275554 -0.141587077
#> 94  094  0.69628171  1.208925132
#> 95  095 -0.13854106  1.258174275
#> 96  096  0.88653021  0.116220578
#> 97  097 -0.65110152 -0.517244407
#> 98  098  1.59289785  1.594487347
#> 99  099  0.65979616  2.169808366
#> 100 100 -1.19656581 -0.660401304