Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5340958
#> Y_2 0.5340958 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5960566 -0.9798449
#> 2  0.9488159 -0.4275248
#> 3  2.2708645  0.1798449
#> 4 -1.0951251  1.9329627
#> 5 -0.7744172 -0.6772435

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.73260484 -0.12601066
#> 2   002  0.31905730 -0.10422854
#> 3   003 -0.77384998  0.39338157
#> 4   004  0.32536454  0.95902753
#> 5   005 -0.84970806 -0.52181717
#> 6   006 -2.49118540 -2.25600517
#> 7   007  0.75171333 -0.48084713
#> 8   008 -1.84012600 -1.89547803
#> 9   009  1.68586299  0.63824806
#> 10  010 -0.23850730 -1.39925492
#> 11  011 -0.15297603 -0.05490544
#> 12  012  0.55783544  0.22670730
#> 13  013  0.38903281 -1.24025879
#> 14  014 -1.30336530 -2.00049778
#> 15  015  1.17440470  1.13858199
#> 16  016  0.34230404  1.35516550
#> 17  017  2.04273747  1.88777259
#> 18  018  0.70786618  0.00787351
#> 19  019 -0.35676287 -0.29335553
#> 20  020  0.05004984  1.08091051
#> 21  021  0.01098254  0.43833320
#> 22  022  1.90178428  1.63432093
#> 23  023  2.13878074  0.23882238
#> 24  024  0.28832345 -0.05161835
#> 25  025 -0.24852313  0.51114252
#> 26  026  0.60922616  1.13181458
#> 27  027  0.27064715 -0.52987309
#> 28  028 -0.57317211 -0.13107903
#> 29  029  1.39227213  0.96838851
#> 30  030 -1.78186838 -1.03543416
#> 31  031 -0.64275199 -0.92005339
#> 32  032  0.26068803  1.04619533
#> 33  033 -0.94438413  0.49376808
#> 34  034  0.82423176  0.01137903
#> 35  035 -0.38217021 -1.10803227
#> 36  036 -1.70187970 -0.41321829
#> 37  037 -1.41179414 -1.09987679
#> 38  038  0.06960392 -0.08363123
#> 39  039 -0.51257693 -0.90230649
#> 40  040  1.15556366  1.82548113
#> 41  041 -1.63158369 -1.69688634
#> 42  042 -1.50595757 -1.57905591
#> 43  043 -1.08679784 -0.94663355
#> 44  044  0.66776971 -0.32177814
#> 45  045 -0.23454169 -0.54211496
#> 46  046 -0.06379981 -0.85203056
#> 47  047  1.48167594  2.69339368
#> 48  048  1.30109761  1.32408353
#> 49  049 -0.78700154 -1.12449722
#> 50  050  1.51765979  0.20747973
#> 51  051 -1.63160757 -1.36654545
#> 52  052  0.82665546  0.41712472
#> 53  053  0.12197178  0.87504686
#> 54  054 -1.76946526 -0.55400485
#> 55  055  0.99074334  1.59517827
#> 56  056 -1.91595482 -1.09786381
#> 57  057 -0.17189710 -0.48130870
#> 58  058  0.24332142  0.30654724
#> 59  059  1.69629413  0.87047720
#> 60  060  0.31169856  1.70501651
#> 61  061 -0.71079777 -0.85063353
#> 62  062 -2.30408342 -2.15794617
#> 63  063  0.58720162  0.51548433
#> 64  064 -0.04815949 -0.74123066
#> 65  065  1.61489713  0.84850807
#> 66  066 -1.37196375 -2.03136832
#> 67  067 -0.40855976  0.55780696
#> 68  068 -1.15531121 -1.50845822
#> 69  069 -1.32454107 -0.74685020
#> 70  070 -0.05976620 -1.87292440
#> 71  071  0.85947153 -0.39043398
#> 72  072  0.18226040 -0.76013959
#> 73  073 -0.14218577 -0.06428133
#> 74  074 -0.05645564 -0.09681400
#> 75  075 -0.18882849 -0.49204728
#> 76  076  1.61934652  1.27779636
#> 77  077 -0.22823898  1.12356606
#> 78  078 -0.36803058 -0.71716899
#> 79  079  0.84529041  0.05489427
#> 80  080 -1.67881786 -0.80439359
#> 81  081 -0.79809781  0.30136342
#> 82  082  0.48618988  1.58179911
#> 83  083 -1.78554412 -2.19409814
#> 84  084 -0.83751262 -0.68180692
#> 85  085 -0.20299934  0.76713182
#> 86  086  1.07999563  0.61952396
#> 87  087 -0.98715576 -1.86657269
#> 88  088  0.97914059  0.41019208
#> 89  089 -2.20505527 -1.96754550
#> 90  090  2.17405982  1.90455289
#> 91  091 -1.24863817 -1.68405517
#> 92  092 -1.62188735 -0.90335622
#> 93  093  1.06618093 -0.06876279
#> 94  094  1.30497534 -0.41588074
#> 95  095 -0.56360908  0.34248268
#> 96  096 -0.71912980 -1.36320739
#> 97  097 -0.47391941  0.45508088
#> 98  098  0.66432083 -1.39521478
#> 99  099  0.35252479 -0.04689415
#> 100 100 -0.25070108 -0.06787333

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  2.643663540  1.11544996
#> 2   002 -0.582951688  0.35377768
#> 3   003  0.738168818  2.15798510
#> 4   004 -0.421105817 -0.03718527
#> 5   005  0.363095951 -1.03483525
#> 6   006 -0.138278634 -2.46974277
#> 7   007 -1.713100866 -2.09025080
#> 8   008 -0.990507695 -0.80571051
#> 9   009 -0.093532959 -1.20851593
#> 10  010  0.517011864 -0.66422571
#> 11  011  0.118642853  0.82214556
#> 12  012 -1.130101259 -0.81586291
#> 13  013  0.018335069  0.18534895
#> 14  014 -0.757990275  1.45257645
#> 15  015 -0.792981532 -1.24372959
#> 16  016 -0.248923638  0.13774366
#> 17  017  0.214427612 -0.44389016
#> 18  018  0.400164604  1.20312852
#> 19  019  1.879402614 -0.36260328
#> 20  020 -0.802497250 -1.02165185
#> 21  021  0.272055198  0.98276272
#> 22  022 -0.464922035 -0.48190215
#> 23  023 -0.392716440  1.32278004
#> 24  024  0.507372996 -0.08048615
#> 25  025  0.607604491 -0.81536173
#> 26  026 -1.499485646 -1.39486416
#> 27  027 -0.797737523  0.32812691
#> 28  028  1.301952940  0.83243677
#> 29  029 -1.815779523 -0.41653873
#> 30  030  1.060768085 -0.03253223
#> 31  031 -0.675942670  0.40777191
#> 32  032  0.009530847 -0.86219658
#> 33  033  0.235818415 -1.20761387
#> 34  034 -0.627387763 -0.59398539
#> 35  035 -0.030058293 -0.53981169
#> 36  036 -0.338878321  1.64834022
#> 37  037  0.072393096 -2.16764923
#> 38  038  0.405993674 -0.59088419
#> 39  039  1.275241165 -0.75251435
#> 40  040  1.035199559  0.20180768
#> 41  041 -0.092369886  0.31454662
#> 42  042  0.265740390 -0.70920261
#> 43  043  0.295056375 -0.21367389
#> 44  044  0.375047553  0.02293478
#> 45  045  1.517334902 -0.86136038
#> 46  046  1.447891460  0.79421685
#> 47  047  1.114468102 -0.39778282
#> 48  048 -0.882038531 -1.03605456
#> 49  049 -0.705925924 -1.13205570
#> 50  050 -0.846597827 -0.32939854
#> 51  051 -0.524790589 -0.80035331
#> 52  052 -0.847984160 -1.22796765
#> 53  053 -0.322488408  1.24447777
#> 54  054  0.792091941  0.14635130
#> 55  055 -0.037651615 -1.90573079
#> 56  056  1.301863497 -0.16966751
#> 57  057  1.280956559  0.61248476
#> 58  058 -1.190942684 -1.31620906
#> 59  059 -0.341982290  1.99749520
#> 60  060  0.845939843  0.47738315
#> 61  061  0.244061252  0.58672542
#> 62  062 -1.514515704 -2.03508778
#> 63  063  0.516196395  0.74362619
#> 64  064  1.564362054  0.27714741
#> 65  065  1.128558778  2.36625230
#> 66  066 -0.151405673 -0.84015828
#> 67  067  0.506059363  0.93900280
#> 68  068 -0.112456504 -0.46594132
#> 69  069  0.115842832 -1.13016262
#> 70  070 -0.977058499 -0.04600715
#> 71  071 -1.535758682  0.39361596
#> 72  072 -0.393827279 -1.71609709
#> 73  073  0.419488311 -0.39576795
#> 74  074  0.287117121  0.38147412
#> 75  075  1.372330274 -0.70930491
#> 76  076  0.425609492 -0.79869983
#> 77  077  0.785804037 -0.79744980
#> 78  078 -0.663801521 -1.34811429
#> 79  079 -2.101016599 -0.03569688
#> 80  080  2.175426469  2.38233780
#> 81  081 -1.621785547 -0.55570405
#> 82  082 -0.201238956 -0.53979080
#> 83  083  1.557946211  0.80503710
#> 84  084  0.644098286  1.35218779
#> 85  085 -0.229788415  0.20703833
#> 86  086  1.608916700  1.82215588
#> 87  087  0.286402353  0.47433673
#> 88  088 -1.009095502 -1.52465213
#> 89  089 -0.302957724 -0.56559165
#> 90  090  0.373813332 -0.77206343
#> 91  091  0.299383110 -1.86327327
#> 92  092 -1.611856812 -1.12231908
#> 93  093  0.300828818  0.42747139
#> 94  094 -0.778282386 -0.13828633
#> 95  095  1.860090026  2.13395870
#> 96  096 -0.104887977  0.77651451
#> 97  097  0.392791258  0.26139383
#> 98  098  0.366395061 -0.11092383
#> 99  099 -1.515916525 -1.01272292
#> 100 100  0.758236255 -0.88927534