Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5083923
#> Y_2 0.5083923 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.4000265 -0.09884022
#> 2  0.8703785 -0.81814618
#> 3 -0.3796293 -0.08620370
#> 4  1.0740083  1.39594006
#> 5  0.2961250  0.85793587

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.07823878  1.09520200
#> 2   002  0.47729520  0.42375397
#> 3   003  0.04979795 -0.37099258
#> 4   004 -0.87174255 -0.45859004
#> 5   005 -0.54522699 -1.22387288
#> 6   006 -0.34338261 -1.71701301
#> 7   007  0.85487739  0.05782238
#> 8   008  0.08315905 -0.57961546
#> 9   009 -0.02436942 -1.04787215
#> 10  010 -2.11695879 -0.72932764
#> 11  011 -0.71055657  0.39294784
#> 12  012  0.41545678 -0.71138833
#> 13  013 -0.43010034 -0.63439710
#> 14  014  0.39473638 -0.68722496
#> 15  015 -0.43058892  0.07449770
#> 16  016 -0.74050819 -0.93663927
#> 17  017  1.54356357  1.24047655
#> 18  018  0.53369900  0.34610817
#> 19  019  0.10683988 -0.28855043
#> 20  020 -0.91599095 -0.73918958
#> 21  021 -0.92808227  0.11212697
#> 22  022 -0.20308985  1.24513426
#> 23  023  0.80010759  0.26549252
#> 24  024  1.55362975  0.26782481
#> 25  025 -0.08458223 -0.04497048
#> 26  026 -0.37326465  1.39728007
#> 27  027  1.75392681  1.57214556
#> 28  028 -0.53914723  1.12421389
#> 29  029 -2.15315251 -2.08578141
#> 30  030  2.32400022  1.25636839
#> 31  031  1.03412542 -0.17965139
#> 32  032  0.93505326 -0.21546609
#> 33  033  0.88695723  0.82641011
#> 34  034  0.94192566  0.46702559
#> 35  035 -0.79991843 -0.29864671
#> 36  036 -2.00140338 -0.63006235
#> 37  037  2.61215773  1.18109139
#> 38  038  0.04181707  0.23245595
#> 39  039  2.22621478  1.83282987
#> 40  040 -1.11585550 -0.71033928
#> 41  041  1.29276074  1.42815044
#> 42  042  0.59815877  0.07154182
#> 43  043 -2.17257747 -1.65095223
#> 44  044  0.33145043  1.18319853
#> 45  045  0.02360102  0.25982833
#> 46  046  0.23048751  0.50409909
#> 47  047 -1.06472980 -2.02176440
#> 48  048 -1.92518884 -1.44844968
#> 49  049 -0.69029913 -1.60735036
#> 50  050  0.24575203 -0.20619004
#> 51  051 -1.64586249  0.27754313
#> 52  052 -0.12584039 -0.89927677
#> 53  053 -0.21418473  1.28553796
#> 54  054 -0.64430034 -0.67624895
#> 55  055 -0.01019817 -0.52223118
#> 56  056  1.39908088  0.07003091
#> 57  057  0.49857027  0.58049677
#> 58  058 -0.28582450  0.38097920
#> 59  059  0.34643819  0.42212956
#> 60  060 -0.96774752  0.10128009
#> 61  061 -0.81137744 -0.79573279
#> 62  062 -0.01435136  0.48867161
#> 63  063  0.72446412 -0.63756168
#> 64  064  0.66215544  1.25189590
#> 65  065  0.19930376  0.02699971
#> 66  066  1.01685264 -0.52495309
#> 67  067  0.17506419 -0.27605804
#> 68  068  0.07284440  0.72421558
#> 69  069  0.31814885  0.77470349
#> 70  070 -1.60577266 -3.04720995
#> 71  071 -1.77306399 -0.96260885
#> 72  072 -0.73896421 -0.68092415
#> 73  073 -0.54372568  0.62364807
#> 74  074  0.44807654 -0.43346558
#> 75  075  0.53206085  1.10849502
#> 76  076 -0.12381942 -0.25491328
#> 77  077 -1.74116146 -1.11823375
#> 78  078  1.28973574  1.46776944
#> 79  079  0.53379771 -0.60564174
#> 80  080  0.65144168  0.85180440
#> 81  081  0.31262338  1.05390107
#> 82  082  0.31280812 -0.16938891
#> 83  083 -0.36619312 -0.29393510
#> 84  084 -0.58595127 -1.04533351
#> 85  085 -1.40934568 -1.47934902
#> 86  086 -1.53168185 -1.99986080
#> 87  087  0.35806224 -0.15276514
#> 88  088  0.84886222  1.94624391
#> 89  089 -0.08223396 -1.30558119
#> 90  090  1.24320218  0.97534807
#> 91  091  0.76257903  0.14347497
#> 92  092  0.31098759  0.44498143
#> 93  093 -1.18207830 -0.38288951
#> 94  094 -0.57671646  0.09325266
#> 95  095 -1.58303818  0.25213604
#> 96  096 -0.36257670  0.77463728
#> 97  097 -0.96106986 -0.58747986
#> 98  098  0.30778098 -1.89129146
#> 99  099  1.18742305  1.31083786
#> 100 100 -1.59718652 -0.11421538

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.006033487  1.38583841
#> 2   002  1.388307281  0.47780802
#> 3   003 -0.461979072  0.09981341
#> 4   004 -0.364204584 -0.91435119
#> 5   005  0.033215593  0.57644556
#> 6   006 -0.496169725 -0.08329626
#> 7   007  0.865811729 -0.56421956
#> 8   008 -0.799177499 -1.76159982
#> 9   009  1.215410456  0.98634975
#> 10  010 -0.474885281 -0.25893370
#> 11  011 -2.245986284  0.71296915
#> 12  012 -0.715462595 -1.11911334
#> 13  013  0.101968481 -0.45786125
#> 14  014  2.039059828  2.03327468
#> 15  015 -0.428918391 -0.66768391
#> 16  016  1.119194063  0.01174650
#> 17  017  1.328465903  2.41976566
#> 18  018 -2.124207935 -0.20506760
#> 19  019 -1.933226506 -1.08825054
#> 20  020  1.464937115  0.29936372
#> 21  021  1.004412690  1.52615195
#> 22  022  0.267919367  0.55417111
#> 23  023  0.851723441 -0.71856065
#> 24  024  1.081846100 -0.25768521
#> 25  025 -1.530244979 -1.32650132
#> 26  026 -0.660058455  0.70843267
#> 27  027 -1.136047172 -0.69277250
#> 28  028  0.149568282  1.10271014
#> 29  029 -0.824352594  0.24625297
#> 30  030 -0.978683239  0.08722864
#> 31  031  1.689578475  2.79000531
#> 32  032 -0.190594567 -2.04497959
#> 33  033  0.420477052  0.62448024
#> 34  034 -2.101876075 -2.17278689
#> 35  035 -2.801312223 -0.71407327
#> 36  036  0.616196556 -0.54977278
#> 37  037 -1.789644695  0.31466542
#> 38  038  1.637543781  0.34408898
#> 39  039  0.689550258  0.78670701
#> 40  040 -1.457290216 -1.58038812
#> 41  041  0.264648675  1.01123971
#> 42  042 -1.799901955 -2.08722825
#> 43  043 -0.197007066  0.29426119
#> 44  044  0.310409796  0.59732971
#> 45  045 -0.963089410 -1.06791134
#> 46  046 -0.517187411 -0.60975450
#> 47  047  0.128652290  0.78874974
#> 48  048 -1.477159892 -0.85832949
#> 49  049 -1.258786425  1.14509032
#> 50  050  1.324764321 -0.29058592
#> 51  051  1.924954297  0.69729413
#> 52  052  0.399578398  1.06950629
#> 53  053 -1.329346731  0.85343562
#> 54  054  0.288290426  1.10623429
#> 55  055 -0.303271879 -0.46039082
#> 56  056  0.259301511  0.35440115
#> 57  057  0.364725074 -0.87722085
#> 58  058  0.925208939  0.02583217
#> 59  059  0.660968256 -0.35681257
#> 60  060 -0.335988479  1.18819850
#> 61  061 -0.057068670  0.69881397
#> 62  062 -0.578101380 -0.55220683
#> 63  063 -0.369762953 -1.24983648
#> 64  064 -0.033742173 -0.19532602
#> 65  065  0.965657359  1.33332685
#> 66  066 -1.032277918 -1.24563055
#> 67  067  0.732591699  0.08404557
#> 68  068  1.064549531  0.33071400
#> 69  069 -0.718905760 -1.26480258
#> 70  070 -2.536884531 -2.42721778
#> 71  071 -1.009955259  0.26702731
#> 72  072 -0.350157080 -0.36505763
#> 73  073  0.555264746 -0.25662952
#> 74  074 -0.162824999  0.05589837
#> 75  075  0.578247567 -1.26731616
#> 76  076 -0.056624431  0.74944426
#> 77  077  1.473453305 -0.46805070
#> 78  078 -1.419832887 -0.45030639
#> 79  079  1.065581662  1.06265933
#> 80  080 -2.271357357 -0.43695070
#> 81  081 -0.322884480  0.22904230
#> 82  082  0.270635233  0.85064400
#> 83  083  2.564765856  1.33370117
#> 84  084  0.624075544  0.76500537
#> 85  085  0.413468838  0.97412075
#> 86  086 -1.127520885 -1.34427290
#> 87  087  2.217048011  2.29446956
#> 88  088  0.011402698  0.11027836
#> 89  089  0.273022271  0.31975032
#> 90  090 -0.379070988 -0.22873403
#> 91  091 -1.658481563  0.24607424
#> 92  092 -0.877486164 -0.51324514
#> 93  093  0.595303502  0.29803425
#> 94  094 -0.261441605 -0.64475219
#> 95  095  0.910497658  1.46076152
#> 96  096 -0.600016891  0.57232429
#> 97  097 -1.453200160 -0.65739555
#> 98  098  1.416523787  0.82717739
#> 99  099  0.678821080  1.10771392
#> 100 100  0.006167344  0.67649495