Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5134624
#> Y_2 0.5134624 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1       Y_2
#> 1 -0.3130913 -1.359308
#> 2  3.0843475  1.325108
#> 3 -0.4060826 -1.468142
#> 4 -1.9040057 -1.767392
#> 5 -1.4432125 -2.125062

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.14108546 -0.661212767
#> 2   002  0.22264466 -0.361722521
#> 3   003  0.44923042  0.211683081
#> 4   004  0.36732055 -1.186140356
#> 5   005  1.24216822  1.233262593
#> 6   006  0.78946784  0.392616525
#> 7   007  1.43197215  0.832490042
#> 8   008 -0.51377032 -1.079230681
#> 9   009 -0.15466163 -0.105285080
#> 10  010  1.55115356 -1.068341967
#> 11  011 -1.04273428 -0.447735597
#> 12  012  1.12611603  1.210698662
#> 13  013  0.24470171  0.277962405
#> 14  014  0.51881457  0.763843144
#> 15  015  1.24084223  0.837244855
#> 16  016  1.16902198 -0.986314028
#> 17  017  1.68435723  0.580831141
#> 18  018 -0.33459203  0.564443266
#> 19  019 -0.31659968 -0.140414049
#> 20  020 -1.19002150 -0.260399762
#> 21  021 -0.69953715 -0.647609651
#> 22  022  1.14945136 -0.250830211
#> 23  023 -1.09086254 -0.854544901
#> 24  024  0.40345178  0.165267340
#> 25  025  0.03504716 -1.066240367
#> 26  026  0.84202371 -0.062257092
#> 27  027  1.34673421  1.611106348
#> 28  028 -0.67046848 -0.820732589
#> 29  029  1.33056279  1.690138663
#> 30  030  1.49443616  0.270335549
#> 31  031 -2.09132505 -1.659487040
#> 32  032 -0.06135926  1.042114660
#> 33  033  0.46427674  1.175905673
#> 34  034  0.79835810  1.467625436
#> 35  035 -0.79460257 -0.341147574
#> 36  036 -0.36302851 -0.884145030
#> 37  037  0.09062268  1.192020548
#> 38  038  0.54954682 -0.356806971
#> 39  039  0.28851210  1.770634041
#> 40  040 -0.77316138 -0.244539661
#> 41  041  0.22711407  0.698956333
#> 42  042  0.40175100 -0.869339910
#> 43  043  2.66649248 -0.570852994
#> 44  044 -0.29134650  0.369185808
#> 45  045 -1.33243232 -1.082850763
#> 46  046 -0.08575659  0.542104330
#> 47  047 -1.27537839 -0.164513478
#> 48  048 -0.84474596 -0.314750419
#> 49  049  0.24012347  0.007155376
#> 50  050 -0.07759965 -1.513834391
#> 51  051  1.54189562 -0.703764159
#> 52  052 -0.14192908  0.448820580
#> 53  053 -2.75463549  0.131592418
#> 54  054 -1.79471088 -1.987354389
#> 55  055  0.93259763 -0.312752048
#> 56  056 -1.32744197 -1.328223230
#> 57  057 -2.30771425 -0.599007755
#> 58  058 -0.72058356  0.534640161
#> 59  059  0.03536312 -0.219332208
#> 60  060  0.15759986  1.431521193
#> 61  061  1.79915796 -1.042532089
#> 62  062  0.14309059 -0.591851230
#> 63  063  2.02889500  1.153818422
#> 64  064  0.89390499  1.160099003
#> 65  065  1.82496215 -0.481426320
#> 66  066  1.48005880  1.113923759
#> 67  067  0.39331190  1.295299046
#> 68  068  0.73684233 -0.080873532
#> 69  069  0.99942091  1.378674382
#> 70  070  1.09032724  0.257526831
#> 71  071 -3.21994968 -1.768978251
#> 72  072 -0.18549281 -1.954417627
#> 73  073  0.18579298  1.146897151
#> 74  074  0.04776198  0.122286603
#> 75  075  1.37453313  0.662844221
#> 76  076  1.00613712  0.895854499
#> 77  077  0.44321999 -0.710300800
#> 78  078  0.78304646  0.056974512
#> 79  079  0.85282996 -0.505794767
#> 80  080 -2.00263593 -0.920925725
#> 81  081 -1.02459441  0.009391843
#> 82  082 -0.87588212 -2.978849362
#> 83  083  0.47991728 -0.331383056
#> 84  084 -0.47644422  0.032417432
#> 85  085  0.90639684  0.104782882
#> 86  086  1.96803753 -0.022612828
#> 87  087  1.18848561  1.081829351
#> 88  088  1.85981049  0.332762565
#> 89  089  0.96096026 -1.168172452
#> 90  090  0.29837003  0.578304394
#> 91  091 -0.64791346 -0.931788325
#> 92  092  0.18005769  0.987971039
#> 93  093 -0.35302009  0.274862584
#> 94  094 -0.65566242 -0.345226477
#> 95  095  0.46728444  0.535612116
#> 96  096 -0.57123382 -1.661678129
#> 97  097  2.25001071  0.749879695
#> 98  098 -2.09522216 -0.987775416
#> 99  099  2.27627560  1.643543192
#> 100 100  1.08059125  1.248858697

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.266519542  0.66586324
#> 2   002  1.442002501  1.68577324
#> 3   003 -1.140241412  0.58377509
#> 4   004 -0.505234916 -0.48065784
#> 5   005  0.383734557  2.12403549
#> 6   006 -0.471576216  1.84860066
#> 7   007  0.263570032 -0.12398184
#> 8   008  1.109416889  0.31981053
#> 9   009 -0.840118964 -0.15787160
#> 10  010  0.167227592  0.94001870
#> 11  011 -0.067827861 -0.16815799
#> 12  012  0.306542764 -0.05270481
#> 13  013 -0.182552086 -1.54871282
#> 14  014 -1.150651462 -0.20623084
#> 15  015 -0.142436838 -1.60245712
#> 16  016 -1.510664454  0.34627466
#> 17  017  0.622111579  0.46163531
#> 18  018 -2.550211223 -0.97932852
#> 19  019 -2.497325233 -2.21339394
#> 20  020  0.551307078  0.81572673
#> 21  021  0.449242632 -0.60713635
#> 22  022 -0.034423267  0.62871105
#> 23  023  0.761053690 -0.55224408
#> 24  024 -0.607511507 -1.82858719
#> 25  025  0.468378060  0.10226292
#> 26  026 -1.429728220  0.21082781
#> 27  027 -0.591638186 -0.76326207
#> 28  028 -0.376579074  0.43151583
#> 29  029  0.788011912  0.79292538
#> 30  030  0.562142145  0.29566269
#> 31  031  0.395396181  1.75820864
#> 32  032 -0.600174066 -1.53870049
#> 33  033 -1.133738818 -2.17992226
#> 34  034 -0.103862135 -0.35429169
#> 35  035 -1.318141448 -0.46620781
#> 36  036 -1.100242134  0.04761766
#> 37  037  0.790582226 -0.13359215
#> 38  038  0.455471733 -0.17233053
#> 39  039  0.712553081  0.11619952
#> 40  040 -0.045268145  1.56863769
#> 41  041 -1.074663809  0.90659758
#> 42  042 -0.717659058 -0.28588088
#> 43  043 -0.473930194 -0.46682836
#> 44  044 -0.414524039  1.90150815
#> 45  045  0.659851463 -0.03769620
#> 46  046 -1.504083632 -0.42202225
#> 47  047 -0.235132929  0.34685778
#> 48  048 -0.109271455 -0.84721491
#> 49  049  0.506042025  0.53481191
#> 50  050 -0.229079786 -0.85348937
#> 51  051  0.421976463 -0.54120341
#> 52  052 -1.772167365  1.53250769
#> 53  053  0.514054526  0.34189156
#> 54  054  0.459297258 -0.69773002
#> 55  055  0.049406998  0.33659312
#> 56  056 -0.846241520 -1.43521934
#> 57  057  1.104778067  1.27865840
#> 58  058  0.702043048  1.05402983
#> 59  059  0.173490301  0.98812804
#> 60  060 -0.544411534 -1.85931088
#> 61  061 -0.882880169 -0.70302375
#> 62  062  0.035452197 -0.28504962
#> 63  063  1.846829421  2.27935219
#> 64  064 -0.529365551 -0.56785486
#> 65  065 -0.424432801 -0.59475004
#> 66  066  0.936319987  1.20626824
#> 67  067 -1.325550054 -0.63209091
#> 68  068 -0.718870989 -0.51746423
#> 69  069 -0.287426002 -1.53652156
#> 70  070 -0.615700094 -0.71369432
#> 71  071  1.503197469 -0.79246870
#> 72  072 -0.232960343 -0.55233490
#> 73  073  0.229739403 -0.06493605
#> 74  074 -0.061846255 -0.10645708
#> 75  075  0.827067894  0.62431633
#> 76  076 -0.778915170  0.42593656
#> 77  077 -0.221778699 -1.47353200
#> 78  078  0.003407226  0.52189046
#> 79  079  0.632524960  1.04654298
#> 80  080 -1.111115339 -0.93699042
#> 81  081 -1.401470637 -0.30340863
#> 82  082  1.414449571 -0.61924857
#> 83  083  1.226369850 -0.59340281
#> 84  084 -0.531473350  1.26896532
#> 85  085  0.672353227 -1.04647818
#> 86  086 -0.339405619  1.32297520
#> 87  087  0.765198586  1.27157421
#> 88  088 -0.241458833  0.49224544
#> 89  089 -0.549489721  0.05749710
#> 90  090  1.202459540  1.59225184
#> 91  091 -0.215804028  0.42004965
#> 92  092 -0.787514131 -0.19590860
#> 93  093 -1.105843126 -1.48738541
#> 94  094  0.761917289  0.46549457
#> 95  095 -0.553155415  0.12650319
#> 96  096  1.499694532  0.45370987
#> 97  097  0.847954012 -0.14111452
#> 98  098  0.664771333 -0.25420265
#> 99  099  0.427978431  1.90291657
#> 100 100  0.336325278 -1.17527062