Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5165822
#> Y_2 0.5165822 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2379608  1.3381292
#> 2  1.0854944  1.0518784
#> 3 -0.8547623  0.2832055
#> 4 -0.5858072 -1.6635806
#> 5 -0.8465847  0.1891483

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.80238729  0.44587637
#> 2   002 -0.63806149 -0.49647333
#> 3   003 -0.23405702 -1.68741198
#> 4   004 -1.98808148  0.69925170
#> 5   005 -0.37044628 -0.44027040
#> 6   006 -1.59889446 -0.86666495
#> 7   007  2.04528659 -0.22475756
#> 8   008  0.72547510  0.64719781
#> 9   009 -1.25179883  0.24784357
#> 10  010 -1.25319368  0.45034792
#> 11  011 -2.24347654 -2.20421517
#> 12  012  0.25234938 -0.04815608
#> 13  013  1.54041654  0.33008319
#> 14  014 -0.32010201 -0.84537209
#> 15  015 -0.28476517  0.02849750
#> 16  016 -0.15472434  1.16535922
#> 17  017  1.56700375  1.38664798
#> 18  018 -0.60716659  0.07780802
#> 19  019  1.40346972 -0.02923155
#> 20  020 -0.45728365  1.04663163
#> 21  021  1.39191201  1.87245225
#> 22  022 -1.14762323 -2.14707639
#> 23  023  0.67773392  0.47462793
#> 24  024  0.15390965  0.06157298
#> 25  025 -0.09332728  0.01151249
#> 26  026 -0.36126716 -0.69968085
#> 27  027 -1.71814289  0.73826582
#> 28  028  0.71488577  1.23775458
#> 29  029  1.03305742  0.67805384
#> 30  030  0.28980840 -0.04505455
#> 31  031  0.68794368  0.32670527
#> 32  032  0.26307257  1.53776451
#> 33  033  0.14339551 -0.24241671
#> 34  034 -0.61441728  0.62678899
#> 35  035 -0.77544983 -0.17100918
#> 36  036 -0.56894803  0.61505006
#> 37  037 -1.79316697 -1.63735068
#> 38  038 -1.36836246 -1.21324925
#> 39  039 -0.87300865  0.20446549
#> 40  040  0.25141425 -0.38899783
#> 41  041 -0.14333115 -1.87170793
#> 42  042  0.37813300  0.86810502
#> 43  043 -0.37906200 -0.31455538
#> 44  044 -1.04759489 -1.50069796
#> 45  045 -1.72006648 -1.04604027
#> 46  046 -1.79702633 -1.88737224
#> 47  047  0.47789568 -0.25121109
#> 48  048  0.64838639 -0.74593316
#> 49  049 -2.29391753 -1.26695899
#> 50  050  1.61691843  0.87651300
#> 51  051 -0.11624270  0.14277067
#> 52  052 -0.20505076  0.57312287
#> 53  053 -0.36524702  0.62256200
#> 54  054  0.89557546  1.32601489
#> 55  055  0.44517641 -0.04596271
#> 56  056 -0.32791510 -0.50420427
#> 57  057 -0.75265935 -0.10708491
#> 58  058 -0.54007065 -0.61699566
#> 59  059  0.07035995  0.25634129
#> 60  060 -0.10382281  0.14042013
#> 61  061 -0.26215941 -0.52591498
#> 62  062  0.37173048 -0.16106199
#> 63  063 -0.74289496 -0.84165150
#> 64  064 -0.59413917 -1.18735462
#> 65  065  0.91602568  0.59523654
#> 66  066  0.27399596  0.29276862
#> 67  067 -0.13045659 -1.12991974
#> 68  068  0.46979159 -0.06254710
#> 69  069 -1.58279073  0.25591850
#> 70  070  0.08446970  0.12356263
#> 71  071  0.39500594 -0.64939320
#> 72  072  1.38473653  1.03980902
#> 73  073  1.48569753  1.95000004
#> 74  074  0.94238915  0.50072736
#> 75  075 -0.73989496 -0.51849804
#> 76  076  0.80495438  2.77070930
#> 77  077 -1.05476668 -0.71860237
#> 78  078 -1.16122624  0.33864278
#> 79  079 -0.91531117  0.44323754
#> 80  080 -1.06588681 -1.57551106
#> 81  081  1.65747677  0.94764238
#> 82  082 -0.09667609  1.20034502
#> 83  083 -0.30865641  0.46268880
#> 84  084  0.64051524 -0.47989457
#> 85  085 -1.22815598 -2.21694670
#> 86  086  1.53330066  2.25481826
#> 87  087 -1.06482038 -0.78072249
#> 88  088  0.04939133 -0.14123898
#> 89  089  0.16895301  0.28475386
#> 90  090  1.17420047 -0.76567816
#> 91  091  0.74922568  0.44728725
#> 92  092  1.35075772 -1.02127218
#> 93  093 -0.69106430 -0.92938294
#> 94  094  0.58216848  0.59698550
#> 95  095  0.61568103  0.99857212
#> 96  096 -0.76409938 -0.65758926
#> 97  097  0.38893023 -0.94944202
#> 98  098 -0.81116968 -1.67851161
#> 99  099  0.51107587  0.46005709
#> 100 100 -0.34669701 -1.48191726

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.21130183 -0.14071782
#> 2   002 -2.08616036 -0.11426105
#> 3   003  0.33183997  0.52390310
#> 4   004 -0.28297645  0.80035133
#> 5   005  0.68111780 -0.01061757
#> 6   006  0.88257193  0.83498102
#> 7   007 -0.06932327 -0.42105188
#> 8   008 -0.20232322  0.30310778
#> 9   009 -0.42264518 -1.33414624
#> 10  010  0.78814296 -1.04109254
#> 11  011  0.86404320  2.47923923
#> 12  012 -0.34117422 -1.05930995
#> 13  013 -0.26258892  0.56540988
#> 14  014  1.28338162  1.30900155
#> 15  015  1.09547593  0.33671324
#> 16  016  0.52447765 -0.89246964
#> 17  017  0.24057064 -1.68218957
#> 18  018 -1.13081566 -0.07846999
#> 19  019 -1.19572238 -1.94917249
#> 20  020 -0.21341339  1.08011149
#> 21  021  1.81580014 -0.25022260
#> 22  022 -2.39489298 -1.14289966
#> 23  023  0.07084466 -1.29926629
#> 24  024  0.05796915 -0.50878747
#> 25  025  0.46681541  0.09996340
#> 26  026  0.04718234  0.98163202
#> 27  027  0.83270187  2.89924630
#> 28  028  0.24202759 -0.14064636
#> 29  029  0.88410163  0.27886918
#> 30  030 -0.86596471 -1.22685644
#> 31  031  1.31902014  0.43792813
#> 32  032  2.77678485  2.39045915
#> 33  033  1.34702451  1.49316026
#> 34  034 -1.23237094  0.27788490
#> 35  035 -0.57814888 -0.01773856
#> 36  036  0.09225305 -2.03605147
#> 37  037 -0.21080387 -1.90503772
#> 38  038 -0.26184039 -0.05134628
#> 39  039  1.73989310  1.83127556
#> 40  040  1.27868142  1.30785549
#> 41  041  1.53001252 -0.07770284
#> 42  042 -1.22422763  0.44284284
#> 43  043  0.59225202 -0.57977482
#> 44  044 -0.05239037  1.57571492
#> 45  045  0.58901385  0.19993244
#> 46  046  1.43935593  0.58569009
#> 47  047 -0.04449972  0.16464310
#> 48  048 -0.61230334  0.14771656
#> 49  049  0.12760052  0.24307198
#> 50  050 -0.47184834 -0.30801392
#> 51  051 -0.17317311  0.64365253
#> 52  052  0.53328173  1.27518571
#> 53  053  0.65364748  0.24685168
#> 54  054  0.27919809  0.31321139
#> 55  055  1.45210938  1.40940275
#> 56  056 -0.08559878  0.49694007
#> 57  057 -1.09414632 -1.73634061
#> 58  058 -0.03921697 -0.42397179
#> 59  059  0.94114259  1.35761861
#> 60  060  0.59864298  1.54410749
#> 61  061  0.37422131  0.37896956
#> 62  062  0.10912651  0.36033873
#> 63  063  0.28407885  1.90608497
#> 64  064 -0.60832257  1.17607561
#> 65  065  0.63818979 -0.88913269
#> 66  066 -0.82698325 -0.77002599
#> 67  067 -1.43389013 -0.68391106
#> 68  068  0.34056504 -0.15606380
#> 69  069 -0.67039331  0.34757455
#> 70  070  0.79001242  0.55156813
#> 71  071  1.01423325  0.30616346
#> 72  072  0.42746828 -1.37584043
#> 73  073  0.31414719 -0.97565528
#> 74  074  1.95261142  2.02895887
#> 75  075  0.65415392  0.04860376
#> 76  076  1.24870422  1.19460965
#> 77  077  0.26999231 -1.25736486
#> 78  078 -0.61683304  0.38893580
#> 79  079  0.80324271 -0.13916874
#> 80  080  0.88799473 -0.82946725
#> 81  081  0.14331932 -0.05863466
#> 82  082 -0.66847412 -0.36526788
#> 83  083 -0.43530250  0.27232555
#> 84  084 -0.39295689 -1.58865513
#> 85  085  0.04703810 -0.81113168
#> 86  086 -0.32349838 -0.67669452
#> 87  087 -0.34481549 -0.92145339
#> 88  088  1.15952746  1.21554367
#> 89  089  1.54669575  2.06293920
#> 90  090  0.01188279  0.65455762
#> 91  091  0.01770338 -0.01682166
#> 92  092 -1.86201844 -1.31326738
#> 93  093  0.48282834  0.13678599
#> 94  094 -0.20504139 -0.47130061
#> 95  095  0.39458161 -0.76922750
#> 96  096 -0.65021583 -0.76462673
#> 97  097 -0.76352883 -0.53884702
#> 98  098 -0.08632569  0.73496146
#> 99  099 -0.84719008 -0.40866217
#> 100 100  1.01639163  2.33502291