Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5168201
#> Y_2 0.5168201 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.3951606  0.1504541
#> 2 -1.0348305 -1.4734736
#> 3 -0.4853945 -0.3938341
#> 4  0.9616082  0.8193118
#> 5  0.5922734  0.1140658

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.125740501  1.977047313
#> 2   002 -0.174487738 -0.662236630
#> 3   003 -1.273346175  0.649077338
#> 4   004 -0.173703491 -0.446832713
#> 5   005  0.885736771 -0.010113173
#> 6   006 -1.054017271 -0.660791193
#> 7   007 -1.098076366 -0.780454100
#> 8   008 -0.423705755  0.038936105
#> 9   009 -0.322860931 -0.137625576
#> 10  010  0.946277967  0.269172729
#> 11  011  0.120968682 -0.482311136
#> 12  012 -0.777351130 -1.503850189
#> 13  013 -2.808262682 -0.772609970
#> 14  014  1.680292441  1.513532555
#> 15  015  1.575707667  0.354969233
#> 16  016  0.277400349  0.926431053
#> 17  017 -0.074921644  1.161734935
#> 18  018 -1.577578277 -1.128106095
#> 19  019  0.495669912  1.573413744
#> 20  020 -0.344603374 -0.460064232
#> 21  021  0.864353621  0.568747976
#> 22  022 -0.856444812  0.326588424
#> 23  023  0.748141827  1.992687528
#> 24  024 -0.313795199  0.139455655
#> 25  025  1.018193289  0.025540348
#> 26  026  0.412100380  0.312396042
#> 27  027 -0.451209099 -0.130969183
#> 28  028 -0.141421060  0.679324992
#> 29  029 -0.115347628 -0.009457271
#> 30  030  0.051298665  1.201611493
#> 31  031 -0.495951536 -0.277547956
#> 32  032 -1.023806067 -1.030934132
#> 33  033  0.041657634  0.132489167
#> 34  034 -0.810169801 -1.567710611
#> 35  035 -0.595476234 -0.537510445
#> 36  036  0.715778425 -0.338986028
#> 37  037 -0.455182224  0.699217136
#> 38  038 -0.170305148  0.003692728
#> 39  039 -0.111861072  0.276917856
#> 40  040  0.413487498  1.168027072
#> 41  041  0.507205608  1.041618057
#> 42  042  0.208004647 -0.811810885
#> 43  043 -0.582231421 -2.176229126
#> 44  044 -1.431535361 -1.114926431
#> 45  045  1.300191216  1.495397387
#> 46  046 -0.355271830 -0.143105823
#> 47  047 -0.146736891 -0.849660308
#> 48  048 -0.119195195  0.731818074
#> 49  049 -0.783798327 -1.346812740
#> 50  050  0.084751150 -0.602774897
#> 51  051  2.054336062  0.666625143
#> 52  052  0.860391745  0.709834915
#> 53  053  1.424319025  0.419961656
#> 54  054  0.529999160  1.695100283
#> 55  055  1.784885865  2.105052556
#> 56  056 -0.568147216  0.698779526
#> 57  057  0.993085695  0.592162971
#> 58  058 -0.140516462  0.625364380
#> 59  059  0.148096964 -0.287234657
#> 60  060 -0.851711417  0.877530259
#> 61  061 -0.633735978  1.119395613
#> 62  062 -0.777796058 -1.645515189
#> 63  063 -0.744674168 -0.663119281
#> 64  064 -0.589865745  0.074795318
#> 65  065 -0.008793382  0.798036516
#> 66  066  0.074959991 -0.111152677
#> 67  067  0.968932923  1.673232639
#> 68  068 -1.093622132 -0.777363993
#> 69  069 -0.124772305 -0.104858517
#> 70  070 -0.393799546 -0.231350027
#> 71  071 -1.753844613 -0.148315007
#> 72  072  0.591079934  0.813830053
#> 73  073 -0.455969523 -0.635469304
#> 74  074  1.170383203  1.526080143
#> 75  075  0.034785613  0.104016723
#> 76  076  0.454503674  0.366471888
#> 77  077  2.029292020  0.298996454
#> 78  078  0.095343444 -1.248387669
#> 79  079  0.252044964  0.121360217
#> 80  080 -0.130166231 -1.556306600
#> 81  081 -2.352259732  0.200510254
#> 82  082  0.604218542 -0.024875608
#> 83  083  0.803835935  0.019354922
#> 84  084  0.662647570  0.287529189
#> 85  085  0.237382719  0.297450911
#> 86  086 -1.402590863 -0.271494783
#> 87  087  0.984581580  2.281411877
#> 88  088 -0.791525450  0.739552180
#> 89  089  0.232099329 -0.642281484
#> 90  090  0.578604590 -0.166761081
#> 91  091 -1.355328475 -1.151193392
#> 92  092 -0.873982195 -1.095445095
#> 93  093 -0.404320940 -1.839132932
#> 94  094 -0.116180264  0.982393656
#> 95  095 -1.264999582 -1.470955799
#> 96  096  0.501494074 -0.602444974
#> 97  097 -1.259697835 -1.540518427
#> 98  098  3.075797088  0.900908752
#> 99  099 -1.039711409 -2.569336576
#> 100 100 -0.566593253  1.965824663

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -0.563456566 -0.384377643
#> 2   002 -0.099430655 -1.087713633
#> 3   003 -0.460451963  0.421838335
#> 4   004 -0.422345016 -1.476495569
#> 5   005  0.149573575 -0.616728381
#> 6   006  0.105430147  1.212342987
#> 7   007 -0.332799125 -0.780262256
#> 8   008 -0.475262542 -0.476054232
#> 9   009 -0.434510648 -1.302829701
#> 10  010 -0.099029958  0.670696077
#> 11  011 -0.296540871 -0.691676898
#> 12  012 -0.206934801 -0.980464643
#> 13  013 -1.056158275 -0.007042807
#> 14  014 -1.922872374 -1.405303846
#> 15  015 -0.664382220 -2.019483104
#> 16  016  1.253800331 -0.474761721
#> 17  017  1.079583009 -0.129240302
#> 18  018 -1.259869360 -0.711084918
#> 19  019 -1.194147775 -0.050463122
#> 20  020  1.933717391  1.952356585
#> 21  021  0.410597815  0.484025660
#> 22  022  0.737602025  0.889779936
#> 23  023  0.588742211 -0.355976727
#> 24  024  1.409599745  0.190589923
#> 25  025  1.305990917 -0.330326778
#> 26  026  1.460853671  0.655346102
#> 27  027  0.333662027  0.900728588
#> 28  028  0.415901521 -0.518656167
#> 29  029  0.148470418  1.506031673
#> 30  030 -0.386026243 -0.908077305
#> 31  031 -0.358039543 -0.866695269
#> 32  032 -1.135117381 -0.393935243
#> 33  033 -0.637729126 -0.170492586
#> 34  034 -0.449679502  0.165648544
#> 35  035 -2.398034406 -1.785408609
#> 36  036  1.241235927  0.862567271
#> 37  037  0.560999444  1.501552072
#> 38  038  0.383167391 -0.589052014
#> 39  039  0.663607863  1.475159497
#> 40  040  0.085267215  1.190635717
#> 41  041 -0.599536186 -1.262830871
#> 42  042  0.065597027  1.035923174
#> 43  043 -0.104772779 -0.853735706
#> 44  044  0.757019745 -0.004551204
#> 45  045  0.969205335  0.781074319
#> 46  046 -0.201073734  0.641320522
#> 47  047 -0.642702468  0.131319903
#> 48  048  0.479499610  0.233447346
#> 49  049 -1.096133769  0.236583811
#> 50  050  1.503959823  1.561867386
#> 51  051 -0.100478868 -1.145327133
#> 52  052  0.549679027  1.214353770
#> 53  053  0.216281509 -0.724054554
#> 54  054 -0.190696912 -0.961685864
#> 55  055  0.546798138  0.078875038
#> 56  056 -0.666148574 -0.996178644
#> 57  057 -0.919277812 -0.606329432
#> 58  058 -0.877668042  1.767416642
#> 59  059  0.126476910  0.137017778
#> 60  060  0.748330068 -0.557748107
#> 61  061 -0.156879315 -0.094217004
#> 62  062 -0.266508688 -0.435346936
#> 63  063 -0.190470166  0.393315439
#> 64  064 -0.307648038 -1.021014079
#> 65  065 -0.288962202 -0.016755157
#> 66  066 -3.071720406 -1.098139051
#> 67  067 -1.433871564 -0.418259811
#> 68  068 -0.090555499 -0.595570056
#> 69  069 -0.246549527  0.024870122
#> 70  070 -0.583388392 -0.388985181
#> 71  071 -0.318127157 -0.237461287
#> 72  072  1.275550936  1.279805497
#> 73  073 -0.561664623  0.608767461
#> 74  074  0.367655390 -1.034288394
#> 75  075  0.229633363 -0.704565102
#> 76  076  0.190973132 -0.193020943
#> 77  077 -0.442303742 -1.086396921
#> 78  078 -3.123718704 -2.101754288
#> 79  079 -0.769173097 -1.603787012
#> 80  080  0.994768221  1.018406278
#> 81  081 -1.538163573  1.048083853
#> 82  082 -0.110621970  0.188415824
#> 83  083 -0.863990954 -0.249627892
#> 84  084 -1.464573774 -1.052617219
#> 85  085 -0.208900641 -1.684253083
#> 86  086 -0.001970765  1.019905649
#> 87  087  0.050342511  0.487314100
#> 88  088  0.588651694  1.750087372
#> 89  089 -0.549330452 -1.201005126
#> 90  090  0.546887761  0.088016782
#> 91  091 -0.044540793 -0.951181116
#> 92  092 -0.059777974 -1.532903026
#> 93  093  1.215984013  0.389814997
#> 94  094 -1.028296988 -0.506810314
#> 95  095  0.100303467  1.074092194
#> 96  096  0.706699477 -0.103924537
#> 97  097 -0.615010386 -0.713903758
#> 98  098  0.476645283  0.745691901
#> 99  099  0.749035024  0.555184087
#> 100 100 -0.132729380  1.730298424