Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4732567
#> Y_2 0.4732567 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.4043672  0.02667607
#> 2 -0.6089451 -0.90769705
#> 3 -0.3470544 -0.30303460
#> 4 -0.4773859 -0.07482654
#> 5 -0.6634644  0.30613170

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.788266470  0.76038922
#> 2   002  0.356563106 -0.64323059
#> 3   003  0.085565660 -0.41903652
#> 4   004 -0.518366603 -1.73367318
#> 5   005 -0.456343499 -0.82830642
#> 6   006  1.063662117  0.46862584
#> 7   007  0.539211531 -1.29281512
#> 8   008 -0.085829898 -0.20796434
#> 9   009 -0.515894331  0.51687890
#> 10  010  0.375365921  2.34995042
#> 11  011  0.688390175  2.20168226
#> 12  012 -0.800353582 -0.96555143
#> 13  013  1.233215542 -0.57963114
#> 14  014  1.881096151  0.86380800
#> 15  015 -0.742574717 -0.26760341
#> 16  016  1.166845695  0.12484254
#> 17  017 -0.580410132 -0.51783310
#> 18  018 -0.169850715 -0.34857813
#> 19  019 -0.404184098 -0.58079306
#> 20  020  1.013814614 -0.16985232
#> 21  021  0.937579445  1.05572249
#> 22  022  0.207228993  1.86938654
#> 23  023  0.551987434  0.45668279
#> 24  024 -0.073502383 -0.03688587
#> 25  025 -1.718236805 -1.26141574
#> 26  026 -0.427210430 -0.18630688
#> 27  027  1.055028034  0.07415330
#> 28  028  0.103037184  0.24515765
#> 29  029  2.083657602  0.98047704
#> 30  030  0.973986274  0.25973153
#> 31  031 -1.104932087 -0.24391167
#> 32  032 -0.214997909 -0.04467367
#> 33  033 -0.841643886 -0.86788402
#> 34  034 -1.455804891 -0.61029817
#> 35  035  0.933309298 -0.54604433
#> 36  036  0.833241398 -0.16438179
#> 37  037  1.276710540 -0.57021793
#> 38  038 -0.247867241  1.48990042
#> 39  039  1.709510609 -0.45471486
#> 40  040  0.489996007 -0.70196894
#> 41  041 -1.934693003 -1.08527552
#> 42  042  0.178437923 -0.15847548
#> 43  043  1.236811007  0.10239247
#> 44  044  0.394965378 -0.21200165
#> 45  045 -1.245669356 -0.20927818
#> 46  046 -0.111089463  0.84063732
#> 47  047 -2.000849470 -2.39487983
#> 48  048 -0.172024470  1.26272548
#> 49  049  0.652925728  1.15329839
#> 50  050  1.992136884  0.79744522
#> 51  051 -1.286970765 -1.15820723
#> 52  052 -1.346672172 -1.27773717
#> 53  053  1.806405472  2.07028381
#> 54  054 -1.111243711 -1.09987263
#> 55  055 -1.521723888 -0.26675796
#> 56  056  0.333140299  0.91502342
#> 57  057 -0.015916013 -1.16617148
#> 58  058  0.472254112  1.39531491
#> 59  059 -2.401415931 -0.43416855
#> 60  060  1.216049952 -0.24132103
#> 61  061  1.557322119  2.39279097
#> 62  062  0.628913939 -0.07713916
#> 63  063  1.074427900 -0.88348309
#> 64  064 -0.297633964  0.27449949
#> 65  065  0.268558243  0.63294786
#> 66  066 -1.138109254 -1.08789790
#> 67  067 -1.143683733 -1.46618753
#> 68  068 -0.300359730 -0.51786464
#> 69  069  1.875979143  0.86630440
#> 70  070  1.087018965  0.65933594
#> 71  071  1.769962990 -0.14967550
#> 72  072  0.091862327  0.06063509
#> 73  073  0.936655047  1.05942239
#> 74  074  0.728394246  2.15339673
#> 75  075 -1.476789686 -0.11133650
#> 76  076 -0.694177687 -0.64319615
#> 77  077  0.689861634  1.18464161
#> 78  078  1.690603369  0.84019905
#> 79  079  1.763973044  0.61839752
#> 80  080 -1.781701506 -1.41112505
#> 81  081  0.141755412  0.45851097
#> 82  082  0.695630300  0.39783816
#> 83  083  0.241948183 -1.24330152
#> 84  084  0.959928645  2.89645029
#> 85  085  0.333752997 -0.25635227
#> 86  086 -0.556620835 -2.54907692
#> 87  087  0.633630326  0.13203878
#> 88  088 -0.119643523 -0.82174632
#> 89  089 -0.154127101 -1.04612544
#> 90  090  0.758128664 -0.24443473
#> 91  091 -1.301849589 -2.05428717
#> 92  092  0.021449950 -0.06036668
#> 93  093  0.007626128 -0.44263668
#> 94  094  0.512561417  1.19722061
#> 95  095 -0.267315874 -1.78918943
#> 96  096 -1.112468317  0.05197427
#> 97  097 -0.275205036  0.80244253
#> 98  098 -1.755759254 -0.63003683
#> 99  099 -2.102198574 -0.69526984
#> 100 100 -0.057910238 -1.03003322

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.313386653  1.13445125
#> 2   002  1.811367583  1.45936624
#> 3   003 -0.704834139 -0.17869643
#> 4   004  1.965673646  0.15648957
#> 5   005  0.796757374 -0.18723277
#> 6   006 -1.277320751  1.06460135
#> 7   007  0.234848722  1.08786606
#> 8   008  1.023736460  0.92254884
#> 9   009 -1.382977566 -1.30189407
#> 10  010  0.886682748  1.08949996
#> 11  011 -0.497324134 -2.37053948
#> 12  012 -1.052726262 -0.81345767
#> 13  013 -0.421111524 -0.32048524
#> 14  014 -0.831529050  0.21446886
#> 15  015 -1.297371510  0.30345143
#> 16  016 -0.039604226  0.10172402
#> 17  017 -0.925759939 -2.17084348
#> 18  018 -1.146691817 -1.31382477
#> 19  019  0.253714462  0.66687021
#> 20  020  0.393326831 -0.71469223
#> 21  021 -1.384574382 -0.47389853
#> 22  022 -0.122978788  0.40561579
#> 23  023  1.470433294  0.05522928
#> 24  024  0.396308427  0.63852170
#> 25  025  0.728863281  1.01730318
#> 26  026 -0.965093965 -0.12364168
#> 27  027 -1.141405630  0.75308326
#> 28  028  0.730151078  0.48667594
#> 29  029 -0.385269012 -0.54208894
#> 30  030  0.410944985  1.26294582
#> 31  031 -0.858234531 -0.17872811
#> 32  032 -0.004372498  0.26783485
#> 33  033 -0.826660539 -0.17854393
#> 34  034  1.135339872  0.49724130
#> 35  035 -0.446434268 -0.68028168
#> 36  036 -0.248375988  0.16075021
#> 37  037  0.042560961  0.59080191
#> 38  038 -0.091701005  0.96150263
#> 39  039  0.644754007  0.05657414
#> 40  040  2.331519818  0.99556103
#> 41  041 -0.849358901  1.46746329
#> 42  042  0.585060779 -0.20791311
#> 43  043  0.943027861  1.89966151
#> 44  044  2.623720008  2.11990916
#> 45  045  0.616696055 -0.36636555
#> 46  046  0.070535130 -0.37580092
#> 47  047 -0.271354514 -0.94611875
#> 48  048 -0.501422562 -1.01213716
#> 49  049  0.286137176  0.06760652
#> 50  050 -0.292012066  0.43877841
#> 51  051  0.988517511  2.26445906
#> 52  052  1.610002230  1.19204101
#> 53  053 -0.277617405 -0.19682410
#> 54  054 -2.478134929  0.44604873
#> 55  055  0.805960727 -0.04827933
#> 56  056  0.967808264  0.09620011
#> 57  057 -1.400507152  0.88520605
#> 58  058 -1.152997948 -1.52642529
#> 59  059 -0.490979210 -0.33299845
#> 60  060 -0.363323877 -1.17929440
#> 61  061 -0.640268295 -0.40980716
#> 62  062 -0.232597472 -1.54152435
#> 63  063  0.251842510  0.11279375
#> 64  064 -0.902100884  0.18548157
#> 65  065  0.196653251  0.65880517
#> 66  066 -1.307307839 -1.16656610
#> 67  067 -1.106444433  0.39257519
#> 68  068 -0.414511271 -0.15236334
#> 69  069 -0.040948133  0.31411530
#> 70  070 -0.154542289 -0.03878533
#> 71  071 -0.352239980 -0.27631288
#> 72  072  0.029817292 -0.56991261
#> 73  073 -1.222721034  0.45008648
#> 74  074  0.043852918 -1.21093822
#> 75  075 -0.742713795 -0.05613092
#> 76  076 -0.471150751 -0.88127403
#> 77  077  1.375312657  1.93287269
#> 78  078  0.117634038 -0.67075653
#> 79  079  0.668882808 -0.46896401
#> 80  080 -1.058240692 -0.58509276
#> 81  081 -1.155526932  0.16462440
#> 82  082 -0.536692560  0.43613697
#> 83  083  0.900882004 -0.64235143
#> 84  084 -0.628305032 -1.74329581
#> 85  085  0.717623332  0.48804276
#> 86  086  0.597324106  1.89935332
#> 87  087 -0.374463325 -0.05980984
#> 88  088  0.410928025  0.19548590
#> 89  089  0.409986839  0.75860762
#> 90  090  0.124216713  0.41595240
#> 91  091 -0.415028397  0.55146589
#> 92  092  1.349988985  1.32946958
#> 93  093  0.507377508 -0.42256914
#> 94  094  1.929200959  0.56067508
#> 95  095 -0.323231056 -0.12399659
#> 96  096  0.028665757 -0.04222392
#> 97  097  0.197069596  0.59005279
#> 98  098  0.765164087 -1.03436226
#> 99  099  2.752542691  1.74570666
#> 100 100 -1.742847350 -0.25159560