Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5414838
#> Y_2 0.5414838 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1 -1.13968008 -1.57861921
#> 2  0.09529403 -0.26105176
#> 3 -0.29816793  0.07066238
#> 4  0.05572971  0.62226712
#> 5 -0.82848060  0.37989077

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.13847655 -0.19371946
#> 2   002 -0.19194370  0.06178686
#> 3   003  2.24612724  1.33005552
#> 4   004  0.55068559 -0.45382256
#> 5   005  1.76289009  1.98239134
#> 6   006 -1.32717304  0.61678050
#> 7   007 -1.13242197 -0.15738222
#> 8   008 -0.42590593 -0.76339695
#> 9   009  1.49608460  1.40605696
#> 10  010 -0.52123205  1.03071503
#> 11  011 -0.52780288 -0.70147260
#> 12  012  1.70680858  0.49813641
#> 13  013  0.21161262  1.89383595
#> 14  014 -1.01415648  1.49504774
#> 15  015 -0.70362354  0.13017082
#> 16  016  0.16580396 -1.11130740
#> 17  017 -1.49631332 -2.32779730
#> 18  018 -0.21492419  0.17299477
#> 19  019  0.37689133 -0.24809051
#> 20  020 -0.80334248  0.73762462
#> 21  021 -0.70972941 -0.97916065
#> 22  022 -1.23916604 -2.53718834
#> 23  023  0.16012234  1.45373392
#> 24  024  0.48962312 -0.39574489
#> 25  025 -0.70281184  1.13639058
#> 26  026  0.85231256  1.86704656
#> 27  027 -0.97526402 -0.67804921
#> 28  028 -0.05908877  0.28052861
#> 29  029 -1.24387589 -0.03092312
#> 30  030  0.38227003  1.39902928
#> 31  031  0.48877085  1.44821757
#> 32  032 -0.72490089  0.46447143
#> 33  033 -1.75452426 -1.07320235
#> 34  034 -1.92402219 -0.61961969
#> 35  035 -0.83535463  0.89782395
#> 36  036 -0.31340190  0.21885904
#> 37  037 -1.62954853 -1.03872155
#> 38  038  1.41255353  2.24561864
#> 39  039  0.91849016  1.69664630
#> 40  040  1.19961457  1.05141103
#> 41  041 -0.76539232 -0.21510307
#> 42  042  0.51661259 -0.44510894
#> 43  043  0.45662196 -0.97011463
#> 44  044  0.94517604 -0.79357854
#> 45  045  1.51821662  0.31775972
#> 46  046 -0.15492199 -0.87819330
#> 47  047  0.36315677  0.36193780
#> 48  048  1.12937603 -0.41541750
#> 49  049 -0.35902768 -0.64018400
#> 50  050 -0.32833837 -0.08536713
#> 51  051 -0.95094102  0.44899922
#> 52  052  0.75386962  0.90500999
#> 53  053  0.36966364  1.10179697
#> 54  054 -1.16413332 -1.19474477
#> 55  055 -0.46759552 -2.19382335
#> 56  056 -0.51099367 -0.65256356
#> 57  057  0.83355947  0.80987745
#> 58  058 -1.66075570 -1.38265332
#> 59  059 -1.41465017 -0.95846902
#> 60  060  0.24436359 -0.52043264
#> 61  061  0.33271399  2.37497077
#> 62  062 -0.24318030  0.32561065
#> 63  063  1.01835818  0.66278458
#> 64  064  1.12993084 -0.54405640
#> 65  065  0.80274108  0.20609050
#> 66  066  1.39119175  0.25960190
#> 67  067  0.70298778 -0.19532569
#> 68  068  1.05415896  0.26313563
#> 69  069  0.79071952 -0.11644853
#> 70  070  0.72613884  0.44700317
#> 71  071 -2.22734721 -1.67588048
#> 72  072  0.77431167  0.28891125
#> 73  073  0.20901799 -0.57764452
#> 74  074  0.33514386 -0.09393831
#> 75  075 -1.22865403 -1.48288952
#> 76  076 -0.01565919 -0.16595634
#> 77  077  1.48106483  0.85540809
#> 78  078  0.66154859  0.34605491
#> 79  079 -0.19814014 -0.40800853
#> 80  080  1.54422122 -0.18061553
#> 81  081 -0.90275358  1.02757834
#> 82  082  2.69183587  1.70007122
#> 83  083  0.88926957  0.45668831
#> 84  084 -1.99363695 -1.34203929
#> 85  085 -0.25019557  0.52352429
#> 86  086  0.77609901  1.34772652
#> 87  087 -0.42078717  0.96518044
#> 88  088  0.22398532  0.83341083
#> 89  089  0.73022209  0.88487182
#> 90  090 -0.58032283 -0.87249113
#> 91  091 -0.64947583 -0.76071672
#> 92  092 -3.32039417 -0.73305151
#> 93  093 -0.06664397  0.79030493
#> 94  094 -0.22811389 -0.81340088
#> 95  095  1.34651315  1.68551208
#> 96  096  0.18140763 -2.15157561
#> 97  097 -0.33390704  0.79550593
#> 98  098  1.72626077  0.38593095
#> 99  099  0.26804682 -0.05858152
#> 100 100  0.70215543  0.77905817

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.885052376  0.95535309
#> 2   002  0.609091384 -0.05286032
#> 3   003  0.852915348  1.52290728
#> 4   004  1.003825511  0.04269288
#> 5   005 -0.844993356 -0.29965826
#> 6   006 -0.116906993 -0.83075228
#> 7   007 -0.725519915 -0.79117408
#> 8   008  1.069374948 -0.41223360
#> 9   009 -1.022554563 -0.65747354
#> 10  010  0.404279346 -0.34064942
#> 11  011 -1.151002455 -0.57117243
#> 12  012  0.370233842  0.81873132
#> 13  013 -0.424242392 -0.47361405
#> 14  014 -0.694527217  0.33913133
#> 15  015  0.118326632  0.23841050
#> 16  016 -0.380727860 -1.95533802
#> 17  017 -0.636569675  0.37199323
#> 18  018 -0.204872653 -0.42749278
#> 19  019  1.680007744  0.87526858
#> 20  020 -0.262171299 -0.30920844
#> 21  021  0.985403989  2.15768592
#> 22  022 -0.242075688 -1.12651460
#> 23  023  0.284543652 -1.18575708
#> 24  024  1.275258486 -0.20958836
#> 25  025 -0.580702618 -1.91784864
#> 26  026 -0.414830224  0.16803754
#> 27  027  1.187506710  1.50021562
#> 28  028  0.460516136  0.69477877
#> 29  029 -0.876066297 -0.78545907
#> 30  030 -2.357294530 -0.40199010
#> 31  031  2.702795006  2.02633275
#> 32  032  1.634048614 -0.48396855
#> 33  033  1.334660867  0.28073153
#> 34  034  0.327752901  0.99793652
#> 35  035  0.688152114 -1.23152049
#> 36  036  0.428542350 -0.26444867
#> 37  037 -2.159466849 -1.05250789
#> 38  038  0.765795017  1.92485589
#> 39  039 -0.085364030 -1.69908538
#> 40  040 -0.007186676  0.04265633
#> 41  041  1.058625742  0.84798891
#> 42  042 -0.729289397  0.85623613
#> 43  043 -2.105260086 -1.69633628
#> 44  044  0.745150381  0.65988342
#> 45  045  0.488612448  0.41300176
#> 46  046  1.199797019  0.61930194
#> 47  047 -0.129402136  0.74220318
#> 48  048  1.987702709  1.09999486
#> 49  049  0.885310166  0.68176682
#> 50  050  0.234637113 -0.89235098
#> 51  051  1.047304182 -1.63480895
#> 52  052 -1.012788480 -0.19357914
#> 53  053  0.114283661  1.62423992
#> 54  054 -0.323315905 -0.48096266
#> 55  055  1.045738024  1.17473693
#> 56  056 -0.652274089 -0.64783215
#> 57  057 -0.395866816  1.17412070
#> 58  058 -0.607688641  0.05121566
#> 59  059  0.598245083  1.54879261
#> 60  060  0.034202233  0.24418049
#> 61  061  1.049104538  1.08260933
#> 62  062 -0.141989203 -1.15867503
#> 63  063  0.267011057  0.63811219
#> 64  064  0.179881550 -0.63520145
#> 65  065  0.816920407  0.16884319
#> 66  066  0.737512264  0.36446872
#> 67  067 -0.140969270 -0.44852755
#> 68  068  0.372219450  0.61853269
#> 69  069 -1.830482865 -0.77348648
#> 70  070  0.482456428 -1.93198763
#> 71  071 -0.159418433 -0.94032290
#> 72  072  0.065951559  0.57479422
#> 73  073 -0.692868018 -0.78831282
#> 74  074  0.946681513  0.58988538
#> 75  075 -0.306225398 -0.87248846
#> 76  076  0.139931086  0.19573666
#> 77  077  0.697282497  1.88599322
#> 78  078 -1.109378935 -1.49135987
#> 79  079 -0.090969200 -0.30366679
#> 80  080 -0.133363570 -0.29525320
#> 81  081  0.762613445  1.63917237
#> 82  082 -0.078933699  0.22042774
#> 83  083 -2.223739026 -2.10614800
#> 84  084 -0.373717158 -1.54722402
#> 85  085 -0.336379714 -0.80367635
#> 86  086 -1.195073878 -1.45286532
#> 87  087 -0.477145587  0.39035059
#> 88  088 -0.355761123  0.88416856
#> 89  089  1.706715023 -1.05876323
#> 90  090 -0.047913138  0.16941524
#> 91  091 -1.876156514 -1.50923268
#> 92  092 -1.413337306 -0.22780270
#> 93  093  0.841692865 -1.00541183
#> 94  094  0.405923058 -0.14940300
#> 95  095  0.111910525 -1.86196925
#> 96  096 -1.688430697 -0.38770598
#> 97  097 -0.356230741  1.15632229
#> 98  098  0.416441512  0.98696662
#> 99  099 -1.089704770 -0.83582448
#> 100 100  0.495077386 -0.66463031