Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5154178
#> Y_2 0.5154178 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -2.1760532  0.12773873
#> 2 -1.8648745  0.42400399
#> 3  0.1997998 -0.09282594
#> 4 -2.1375786  0.17460238
#> 5 -1.1808881 -0.90496645

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.631166319  0.08915655
#> 2   002 -1.268976686 -2.16043173
#> 3   003  0.751325176  1.31083089
#> 4   004  0.722871394  1.29249292
#> 5   005 -1.385374554 -0.92917016
#> 6   006 -0.443468376 -0.23102784
#> 7   007  0.503058666 -1.03059081
#> 8   008  2.034394164  1.96406638
#> 9   009 -0.339665102 -0.82478332
#> 10  010  1.650947790  0.58245753
#> 11  011 -0.679897558  0.55304070
#> 12  012  0.498071494 -0.38603363
#> 13  013 -0.731486890 -0.30760293
#> 14  014  1.273490828  0.65593701
#> 15  015  0.145947316 -0.20603372
#> 16  016 -0.389267168 -0.91072949
#> 17  017 -0.059239117 -1.07834470
#> 18  018 -0.939801979  0.14175875
#> 19  019  0.684767093 -0.69330347
#> 20  020  0.299206839  0.91366489
#> 21  021 -0.768383668 -0.01555315
#> 22  022  0.283008849 -0.79283769
#> 23  023 -0.001837269  1.04932004
#> 24  024  0.297576060 -0.01304542
#> 25  025  1.939660046  1.65959171
#> 26  026  1.419819530  1.01961225
#> 27  027 -0.401866891 -0.54688731
#> 28  028  0.046416762 -1.16511990
#> 29  029 -0.354450642 -0.72829429
#> 30  030  0.019874458  0.83340531
#> 31  031 -0.246469611  1.89408724
#> 32  032 -0.250430622  0.23915149
#> 33  033 -1.165293502 -0.33517928
#> 34  034 -0.501643264  0.59565408
#> 35  035 -0.284586850  0.22687421
#> 36  036 -1.388646502 -0.28213949
#> 37  037 -0.345929563 -1.57455650
#> 38  038 -0.377642539  0.45245737
#> 39  039  1.560419183  1.77917204
#> 40  040  0.625479863  1.13328213
#> 41  041  0.826070766  1.60463208
#> 42  042  0.703246867  0.27080102
#> 43  043 -1.993588445 -0.62104887
#> 44  044 -0.515137606 -0.93720937
#> 45  045  0.120910054 -0.01492520
#> 46  046  0.397747263  1.38755452
#> 47  047  1.678519793  1.04128488
#> 48  048 -0.415490601  0.60450967
#> 49  049  0.266915072 -0.11371797
#> 50  050 -0.352805795 -2.56652344
#> 51  051  0.354820655  1.04358583
#> 52  052  0.184877028 -0.40734522
#> 53  053 -0.216526062  0.12202801
#> 54  054 -1.224851447 -0.38119407
#> 55  055 -0.025600177 -0.41810217
#> 56  056 -0.184311520 -0.17130330
#> 57  057  0.534379734  0.28877052
#> 58  058 -0.133510711  0.44503781
#> 59  059  0.851604126  0.18396240
#> 60  060  0.242221981  0.28313790
#> 61  061 -0.730805585 -0.85282936
#> 62  062  0.326426542  0.54060301
#> 63  063  0.542284979  1.48307105
#> 64  064  0.164529404 -1.11287426
#> 65  065 -1.227071391 -1.78990145
#> 66  066  0.805096865  0.57997759
#> 67  067 -0.642170563 -0.84394633
#> 68  068  1.197197002  1.03940166
#> 69  069 -1.180573225  0.36502145
#> 70  070 -0.251787391  0.12305150
#> 71  071 -1.526467100 -0.14449096
#> 72  072  0.598286899 -0.51123270
#> 73  073 -2.055681518 -0.81554407
#> 74  074 -0.939018143  0.17317451
#> 75  075 -0.844483504 -0.33334785
#> 76  076  0.755435705  0.26063123
#> 77  077  0.850684033  1.18474654
#> 78  078  1.390335347  0.72604525
#> 79  079 -0.387617534 -1.36076134
#> 80  080  0.130403616  0.08533008
#> 81  081 -0.357732455 -0.50276071
#> 82  082  1.301061569  0.76846170
#> 83  083  1.340353494  1.27052609
#> 84  084 -0.972924674 -0.60074220
#> 85  085  0.763326221  0.95126739
#> 86  086 -1.682735645 -0.77918066
#> 87  087 -2.545067605 -2.34590781
#> 88  088 -0.149125713 -0.55382368
#> 89  089  0.194646135  1.46167927
#> 90  090 -0.012762433 -2.10292620
#> 91  091 -1.440079010 -2.55784065
#> 92  092 -0.420970979 -0.64078920
#> 93  093 -0.284713879 -1.67194362
#> 94  094 -1.000538419 -0.78052027
#> 95  095 -1.217869289  0.10571799
#> 96  096  0.092118019 -0.27719194
#> 97  097 -0.034595153 -1.22009898
#> 98  098  1.337042793  0.91006386
#> 99  099  0.678019654 -0.89458772
#> 100 100 -0.795511666 -0.17485013

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.402623489 -0.87953588
#> 2   002  0.634460158  1.43092100
#> 3   003 -1.412523451 -1.80539491
#> 4   004 -0.390362365  0.03864445
#> 5   005  0.098250473 -0.64733260
#> 6   006  0.476152758  0.93603701
#> 7   007 -0.618126462  0.25576831
#> 8   008 -1.205488722 -1.40587238
#> 9   009  0.559040060  0.61266890
#> 10  010 -0.062724578 -0.06250060
#> 11  011 -1.074402314 -0.80915985
#> 12  012  0.558879347 -0.04127479
#> 13  013 -1.209784544  0.19604132
#> 14  014  0.352855589 -0.83285442
#> 15  015  1.579437457  1.10690761
#> 16  016 -0.275781411  0.80640389
#> 17  017  0.722916431  0.32412902
#> 18  018 -0.616261117 -0.56146892
#> 19  019  0.001955422 -0.35331038
#> 20  020 -0.492986099 -0.74252372
#> 21  021 -0.694566560 -0.36757717
#> 22  022 -1.202749124 -0.87898367
#> 23  023 -0.803208354 -2.30641522
#> 24  024 -0.455807963 -0.28434977
#> 25  025  0.767640169 -1.52535559
#> 26  026 -0.238444707 -1.29424582
#> 27  027 -0.699095337 -1.05506121
#> 28  028  0.107975069 -0.41083201
#> 29  029  3.123184402  1.11277881
#> 30  030  0.614894846  0.68174322
#> 31  031  1.132315761  2.49861468
#> 32  032  0.503656581  1.84525468
#> 33  033  0.183516987 -0.63354997
#> 34  034  0.423766956  1.02888585
#> 35  035  1.053969305  2.13240183
#> 36  036  0.697697812  2.44009563
#> 37  037  0.125592734  0.08370572
#> 38  038 -0.987648859 -0.58920254
#> 39  039 -1.641394714 -1.63052768
#> 40  040 -0.283542911 -0.69161595
#> 41  041 -0.121932454  0.40857038
#> 42  042 -0.410494120  0.42202573
#> 43  043  0.324567653 -0.27237424
#> 44  044 -1.583814629 -1.45041788
#> 45  045  0.677555363  1.77575097
#> 46  046 -0.265638875  0.76733562
#> 47  047  0.289137807 -0.05442382
#> 48  048  1.128560600  0.73419654
#> 49  049  1.745367106 -0.42102007
#> 50  050  1.580186116  1.44436736
#> 51  051 -0.454624465  1.41942945
#> 52  052  2.517246995  0.91492204
#> 53  053 -0.450235206  0.60752826
#> 54  054  0.578513623 -0.34215849
#> 55  055 -0.512206982  0.34745560
#> 56  056  0.877185514  0.20909785
#> 57  057  0.885507017 -0.12110465
#> 58  058 -0.388883270  0.54576533
#> 59  059  0.212082060  1.20026212
#> 60  060  0.464369458 -1.03596664
#> 61  061  0.905511473  1.59641888
#> 62  062 -0.219263139  0.29532676
#> 63  063 -0.917954053  0.41142034
#> 64  064 -0.697027948 -0.56283361
#> 65  065 -0.096861914  0.22664243
#> 66  066  0.646686057 -0.84042609
#> 67  067 -0.340211725  0.14977624
#> 68  068 -0.993310735 -0.22167927
#> 69  069 -1.463546542  0.08806915
#> 70  070  0.577802897 -0.48392454
#> 71  071 -0.573963243 -1.97812335
#> 72  072  0.904000941  1.30305991
#> 73  073 -0.687662653 -0.05261419
#> 74  074 -0.298500445 -0.13368299
#> 75  075  0.404491053  0.13826389
#> 76  076 -0.109185251 -0.23416526
#> 77  077 -1.100185783  0.05361242
#> 78  078 -1.015063216 -1.02010373
#> 79  079  1.486350696 -0.96925395
#> 80  080 -1.383355047 -0.65568259
#> 81  081 -0.314218752 -0.07819199
#> 82  082  1.202121591  2.18701621
#> 83  083  2.641304428  0.35911122
#> 84  084 -0.925598553  0.54947587
#> 85  085  0.875468031  1.93665307
#> 86  086 -0.777987345  0.70743329
#> 87  087 -1.281023099  0.25523625
#> 88  088 -0.958909435 -0.96214318
#> 89  089 -0.708733283 -0.06837171
#> 90  090 -0.440224133 -0.14406803
#> 91  091  2.585018166  0.12886601
#> 92  092 -0.504452666  0.72972082
#> 93  093  0.157085496  0.38215573
#> 94  094 -1.411682659 -0.04469178
#> 95  095  0.090965609  1.38213841
#> 96  096 -0.444905499  0.14297895
#> 97  097  0.569410943 -0.24485417
#> 98  098  0.192896417 -0.85176152
#> 99  099 -0.734926063  0.69199892
#> 100 100  0.009811210  0.30426754