Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4528657
#> Y_2 0.4528657 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1        Y_2
#> 1 0.7623566 1.16216223
#> 2 1.0030543 1.59315201
#> 3 0.2367298 0.01765538
#> 4 0.4546053 0.54908497
#> 5 2.6777823 2.30170161

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.153323831 -0.616299451
#> 2   002  0.156465289  0.982567161
#> 3   003 -0.430101936  0.655384753
#> 4   004 -0.528971932 -0.366001534
#> 5   005 -0.718975158  1.614119410
#> 6   006  0.057661698 -0.086434119
#> 7   007  0.116368008  0.258735563
#> 8   008  0.421986063  1.585207467
#> 9   009 -0.656929999 -1.882630697
#> 10  010  0.190615574  0.982285567
#> 11  011  0.888014553  0.480014404
#> 12  012  0.120649778  0.465211722
#> 13  013 -0.416708032  2.732693798
#> 14  014 -0.026499306 -1.154152744
#> 15  015  0.316181374 -0.632110014
#> 16  016  0.995071945 -0.112456467
#> 17  017 -1.044482689 -0.592063877
#> 18  018 -0.332472429 -1.219057567
#> 19  019 -0.087184511  0.337361212
#> 20  020 -2.134616076 -1.083766291
#> 21  021 -1.062565469 -2.216612350
#> 22  022 -0.435829945 -0.985362384
#> 23  023  0.451326171 -0.102510722
#> 24  024 -0.144781428  0.705699035
#> 25  025  0.118337102 -0.268676032
#> 26  026  0.923052753 -0.044428990
#> 27  027  1.262652941  0.882329565
#> 28  028 -0.038040931 -0.312463646
#> 29  029 -0.596821307  0.478502279
#> 30  030 -0.676966922 -0.824067600
#> 31  031  1.476006182  2.273445601
#> 32  032 -0.695358475  0.102938759
#> 33  033 -0.908546241 -1.222620182
#> 34  034 -2.093672274 -2.794520588
#> 35  035  0.170441706  0.157033756
#> 36  036 -0.760899032  0.098397160
#> 37  037 -0.154204018 -0.865360962
#> 38  038  1.764895041  1.960484493
#> 39  039  0.838773535  1.279228065
#> 40  040 -1.107361863  1.268923430
#> 41  041 -1.213031016  0.003088372
#> 42  042  0.193550825  2.000399209
#> 43  043  0.979635420  0.465341028
#> 44  044  0.075151134 -0.798100160
#> 45  045 -1.973534128 -1.460408583
#> 46  046 -0.976795053 -1.784114150
#> 47  047  1.057878210  1.820044818
#> 48  048 -0.516539244 -0.360827157
#> 49  049 -0.019931988  0.456331120
#> 50  050  0.068814617  0.382828430
#> 51  051  0.678726864  0.939372779
#> 52  052  0.825661880  0.859014332
#> 53  053  0.916850013 -0.440273289
#> 54  054  1.064365406 -1.823805681
#> 55  055 -1.630639807 -1.198123553
#> 56  056  0.889219037  0.499599047
#> 57  057  0.488860277 -0.226753059
#> 58  058 -1.209021239 -0.231453517
#> 59  059  0.387358618 -1.012099232
#> 60  060  0.194590839 -0.081463566
#> 61  061 -0.002424704  0.486838956
#> 62  062  1.534612254  0.981141937
#> 63  063  0.474353775  0.940007633
#> 64  064 -0.050090785  0.590519890
#> 65  065 -0.226232720 -0.864310291
#> 66  066 -0.261634234  1.111919683
#> 67  067 -0.284067314 -0.614802845
#> 68  068 -1.040756457 -1.013178707
#> 69  069  0.367503029  0.212976446
#> 70  070 -1.207477930 -0.875225483
#> 71  071  0.344353143  1.457651105
#> 72  072 -0.179467325  0.625736275
#> 73  073 -0.467649180 -0.100582240
#> 74  074 -1.404975777 -1.713427028
#> 75  075  0.864823924 -0.075538500
#> 76  076 -0.477610275 -0.212119142
#> 77  077  1.189505435  2.552421756
#> 78  078  1.462516806 -0.458956942
#> 79  079  0.643092985 -1.608410579
#> 80  080  0.664001570 -0.526557734
#> 81  081  1.091963250  1.814861269
#> 82  082 -0.777804765 -2.034887932
#> 83  083 -0.587412578 -0.147856699
#> 84  084  0.918452153  0.440783028
#> 85  085 -0.783487235 -0.516389812
#> 86  086 -0.037646199  1.370491379
#> 87  087  0.087875470  0.557303041
#> 88  088  0.592620922  0.740710593
#> 89  089  0.151641033  0.047416046
#> 90  090  0.896774160 -2.270694606
#> 91  091 -1.199693750 -0.338685494
#> 92  092 -1.732740136 -2.451311003
#> 93  093  0.272260799 -0.351030972
#> 94  094 -0.393606936 -0.774595248
#> 95  095 -0.562796741  0.077675827
#> 96  096 -0.762350947 -1.502832198
#> 97  097 -0.979673814  1.170093314
#> 98  098  1.531782887  1.597166721
#> 99  099  1.524467936  0.265471491
#> 100 100  1.812201604  1.111181749

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.058690184  0.901626408
#> 2   002 -0.385906901 -0.549558853
#> 3   003 -0.185534676 -0.017824714
#> 4   004  0.201072153 -0.303926318
#> 5   005  0.660456954  0.448969594
#> 6   006 -1.382496288 -0.753564907
#> 7   007 -0.811219251 -1.894840895
#> 8   008 -1.247493169 -0.766239298
#> 9   009 -0.921775881  1.202820746
#> 10  010  0.659063629  1.193087042
#> 11  011 -0.918787880 -0.212941117
#> 12  012 -1.740914274 -0.599039002
#> 13  013 -0.655033504 -0.227545452
#> 14  014 -0.404794897  0.149292422
#> 15  015  0.657928366  1.907493768
#> 16  016 -1.523695665  0.471053045
#> 17  017 -0.628456509  0.501608411
#> 18  018 -0.303091970 -0.016716640
#> 19  019  0.449401351  0.167534315
#> 20  020  1.745568496  0.179775019
#> 21  021  1.405712195  1.200155498
#> 22  022  0.164754230  1.791477002
#> 23  023 -1.004855140 -0.839247301
#> 24  024 -0.669828336 -0.505680403
#> 25  025 -1.433370272 -0.987835653
#> 26  026  3.902102082  1.509386400
#> 27  027  0.841314663  0.671379071
#> 28  028  0.794919538  1.469894917
#> 29  029 -0.564159887  0.015568683
#> 30  030 -0.109549604  0.054586822
#> 31  031 -0.076384734  0.986540343
#> 32  032 -0.543482994 -0.164869432
#> 33  033 -0.340939885 -0.669955716
#> 34  034 -0.478730216 -1.710492444
#> 35  035 -1.125968029  1.083803475
#> 36  036  0.848503628 -0.244963298
#> 37  037 -0.627555536 -0.321686394
#> 38  038 -0.065801059  0.207630782
#> 39  039  0.616189491  1.258344727
#> 40  040 -0.119721616  1.488430970
#> 41  041 -1.434941906 -0.767172572
#> 42  042  1.384450211 -0.369311626
#> 43  043 -1.221011634 -0.451804549
#> 44  044  1.032516342 -0.097397979
#> 45  045  0.251360374  0.214479691
#> 46  046  1.488713639  1.561225042
#> 47  047  0.170393438  0.003351981
#> 48  048  0.373461920  0.039148428
#> 49  049  0.065335957  1.819950881
#> 50  050  1.251871015  1.265778034
#> 51  051  1.207159969  1.109517616
#> 52  052  0.448883661  1.263123109
#> 53  053 -0.759789313  0.958115091
#> 54  054  1.506337631  1.220525352
#> 55  055 -0.457029209  0.460665550
#> 56  056 -0.224390597 -0.195070176
#> 57  057  0.400697357 -0.223857683
#> 58  058  0.794312677  0.389612105
#> 59  059 -1.759706826 -0.253195778
#> 60  060  0.465731063  0.103511602
#> 61  061 -0.262252355 -0.675569369
#> 62  062 -0.541331634 -1.606896086
#> 63  063  0.386559374  0.564035422
#> 64  064  0.543920798  0.635398242
#> 65  065 -1.600492875 -1.505955732
#> 66  066  1.818417971  3.008575833
#> 67  067  0.298874050  1.744724097
#> 68  068 -1.136505445 -0.879211147
#> 69  069  0.070976752  0.395442066
#> 70  070 -0.363729963 -1.008781999
#> 71  071  0.545938833  0.396621748
#> 72  072 -0.110845928  0.409235895
#> 73  073 -0.382795642 -0.361285835
#> 74  074  0.680899804  0.440323549
#> 75  075 -0.015946282  0.084292296
#> 76  076 -0.916232753 -1.390688746
#> 77  077  0.304590060  2.015173188
#> 78  078 -0.635653419 -1.549096766
#> 79  079  0.367152544  0.326864186
#> 80  080  1.700681752  1.104776135
#> 81  081 -0.008743545  0.219717304
#> 82  082 -0.305382579  0.438501789
#> 83  083  2.424931191  1.549401567
#> 84  084  1.088205513 -0.263502993
#> 85  085  0.629515825  0.195871219
#> 86  086  2.269091912  1.872569972
#> 87  087  0.850938746 -0.190243239
#> 88  088  1.218185354 -0.157611897
#> 89  089  0.061514225  0.519547764
#> 90  090 -0.125031790  1.329192978
#> 91  091 -1.022850952 -0.428457281
#> 92  092 -0.099790902 -0.995719969
#> 93  093  0.732958762 -0.447314119
#> 94  094 -0.107093396  0.709848095
#> 95  095 -1.679507686 -1.935647769
#> 96  096  0.454336269  0.664469814
#> 97  097 -0.780107951 -0.649309136
#> 98  098  0.166512331 -0.369404954
#> 99  099  0.809552284  0.627746400
#> 100 100  1.652895307  1.287706400