Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5069758
#> Y_2 0.5069758 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.4749055  2.5957450
#> 2 -1.2517861 -3.2282485
#> 3  0.7837236 -0.8731466
#> 4  0.5408523 -0.6736938
#> 5  1.5329484  1.3529335

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.323645729 -0.886316282
#> 2   002  0.288684492  0.428870121
#> 3   003 -0.661060451 -1.473422950
#> 4   004 -1.492174442  0.745613517
#> 5   005 -0.654835488 -0.448709613
#> 6   006  0.597992621  0.190165981
#> 7   007  0.896983141  2.118519839
#> 8   008  1.493445509 -0.004127255
#> 9   009 -0.309425456 -0.096552799
#> 10  010  1.236071238  1.817982927
#> 11  011 -1.281005381 -0.388479112
#> 12  012 -0.587131274  0.738702111
#> 13  013  0.323344117 -0.457668045
#> 14  014 -1.147902621  0.208431215
#> 15  015  0.335751510  1.208221556
#> 16  016 -0.835487420 -0.926552632
#> 17  017  0.901686744  1.154845287
#> 18  018 -2.747605468 -1.165591086
#> 19  019 -0.070119106 -0.452688777
#> 20  020 -0.124826321  0.712499949
#> 21  021  0.213207178  0.622148169
#> 22  022 -0.631869978 -1.226851349
#> 23  023  0.086902817 -0.079647989
#> 24  024 -1.411371679 -1.447510595
#> 25  025 -0.720570382 -1.207477914
#> 26  026  1.064388413  0.150383919
#> 27  027  0.846170216 -0.173500026
#> 28  028 -2.173830429 -2.739548655
#> 29  029  0.953806268  0.765034603
#> 30  030  1.150153365  1.952178886
#> 31  031  0.697283427  0.695221117
#> 32  032  1.667797435 -0.395314075
#> 33  033 -0.643172839 -1.598304402
#> 34  034  1.527650513  1.113828441
#> 35  035  0.692841077 -0.250678527
#> 36  036  0.670426308  0.804181188
#> 37  037 -0.478600056 -1.535288095
#> 38  038 -0.380083311 -0.500652061
#> 39  039  1.116945971  0.942687230
#> 40  040 -0.972430908 -0.024418770
#> 41  041  1.113059802 -0.460677532
#> 42  042 -0.142648183 -1.469140822
#> 43  043 -1.092832902  0.487598126
#> 44  044 -0.181569173 -0.562526390
#> 45  045 -0.138613624 -0.607216059
#> 46  046 -0.564880048  0.288348468
#> 47  047 -0.400812431 -0.240584528
#> 48  048  0.309932505  0.107887832
#> 49  049 -0.162845824  0.441542814
#> 50  050 -0.321913734 -1.075958707
#> 51  051  0.059373856  2.240086679
#> 52  052  1.877713398  1.384004874
#> 53  053 -0.449558146 -1.528390269
#> 54  054 -0.607298244  0.772084665
#> 55  055 -0.093284171 -0.125094889
#> 56  056 -0.543084150  0.651807434
#> 57  057 -0.630800272  0.331415260
#> 58  058  0.527500763  0.427425359
#> 59  059 -0.207132311  0.112862751
#> 60  060  0.825217722  0.848339671
#> 61  061 -0.722994259 -0.618660383
#> 62  062 -0.583289678 -0.196269477
#> 63  063  0.961019341  0.506112081
#> 64  064 -2.033001099 -3.182385885
#> 65  065 -0.002691117  0.269557130
#> 66  066  0.785189692  0.615668664
#> 67  067  0.490448835  0.774150464
#> 68  068 -1.741282104 -0.631012164
#> 69  069 -0.049730070  0.288391696
#> 70  070  0.385139908 -0.463670967
#> 71  071 -1.624930388 -1.185992057
#> 72  072 -0.969389656  0.416059629
#> 73  073 -0.559015054 -1.744784039
#> 74  074  1.191532917  2.611808350
#> 75  075  0.086464271  1.252808414
#> 76  076  0.201947548 -0.113075125
#> 77  077  0.870738628  0.678600101
#> 78  078  0.509975895 -0.068265423
#> 79  079 -1.293790065  0.859343283
#> 80  080 -1.929279064 -0.545614422
#> 81  081  0.049027635 -0.781668903
#> 82  082  0.249882785  1.262305378
#> 83  083  0.873018682  0.075062240
#> 84  084 -0.327914221 -0.369453138
#> 85  085  0.341088477  1.347294764
#> 86  086  0.029755723 -1.407945579
#> 87  087 -0.552668185 -0.761245565
#> 88  088 -0.306206617 -0.375284781
#> 89  089 -1.574718601 -2.652724902
#> 90  090  0.490921574  0.353861131
#> 91  091  2.154532239 -1.183033768
#> 92  092  0.149554774 -0.013223679
#> 93  093  0.153410720  0.530118102
#> 94  094 -0.397636988  0.782166466
#> 95  095 -1.304177021 -0.623101003
#> 96  096  1.756173243 -0.371183545
#> 97  097 -1.728984061 -0.467268731
#> 98  098  0.437127083  0.388488575
#> 99  099 -0.919476467 -1.224264441
#> 100 100  0.773626618  1.775682471

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.47070436 -1.797350077
#> 2   002 -0.18409299 -0.128976131
#> 3   003  1.29659597 -0.380325994
#> 4   004  1.86807071  1.039493658
#> 5   005 -0.30947754  0.292698568
#> 6   006  0.88884974  1.472182145
#> 7   007  0.88200226  0.875091072
#> 8   008 -0.86926762  0.098961362
#> 9   009 -0.02350569 -0.026256964
#> 10  010  1.03729483 -0.178094444
#> 11  011  0.76139557 -0.135246493
#> 12  012 -0.17867829  1.068245399
#> 13  013 -2.04009349 -0.361107884
#> 14  014 -1.68439773 -2.687890076
#> 15  015 -0.17421943  0.162055286
#> 16  016  0.18258097  0.255219225
#> 17  017  0.96372757 -0.608782307
#> 18  018  0.42067940 -0.044267550
#> 19  019 -0.32047949 -0.231187624
#> 20  020  1.32307749  0.861669092
#> 21  021  0.33443448 -1.606972085
#> 22  022  1.88223467  1.655318671
#> 23  023  0.06509201 -0.330214904
#> 24  024  0.72057608  0.877249484
#> 25  025 -0.07894358  0.224395391
#> 26  026 -0.04124011 -0.004011396
#> 27  027  2.39769129  1.175856446
#> 28  028 -0.28731986 -0.218095480
#> 29  029 -0.49080498  0.926834906
#> 30  030  1.28960407  0.475833350
#> 31  031  0.43145112  0.215324132
#> 32  032  0.48394966  1.430731822
#> 33  033  0.21658475  1.180484941
#> 34  034  0.03275621  0.726655868
#> 35  035 -0.49559420  0.257273391
#> 36  036  0.65206363 -0.219100667
#> 37  037  0.66313864  0.054938560
#> 38  038  1.30915107 -0.740892759
#> 39  039 -0.96432847 -0.413101683
#> 40  040 -1.08523609  0.116472641
#> 41  041  0.99944454  0.244638723
#> 42  042  1.02439564  1.507879912
#> 43  043 -0.16772978 -0.684191482
#> 44  044  0.55043618  1.749235922
#> 45  045 -0.09354734  0.558637438
#> 46  046  1.69761520  1.053487052
#> 47  047 -0.33688894  1.160540523
#> 48  048 -0.81855036  0.119403537
#> 49  049  1.41982299  0.715133825
#> 50  050 -1.52879503 -1.753741288
#> 51  051  0.89582364  0.665780370
#> 52  052  0.42599504  1.496088665
#> 53  053 -0.60936608 -1.293844535
#> 54  054 -0.92374275 -0.383889131
#> 55  055 -1.38357890 -1.881404027
#> 56  056 -1.74019796  0.424476564
#> 57  057  0.60404203  1.205076440
#> 58  058  1.29528195  2.359549921
#> 59  059 -0.42830219  0.350148481
#> 60  060 -0.14819237 -0.496875864
#> 61  061 -1.12712350 -0.309806913
#> 62  062  1.80658362  0.747502092
#> 63  063  1.65032030  0.747347687
#> 64  064 -0.34620345  0.004348249
#> 65  065 -0.81091600  0.425857323
#> 66  066 -0.05385149 -1.203611100
#> 67  067 -0.88884412 -0.870799763
#> 68  068 -2.06220045 -0.626069350
#> 69  069  1.00483146 -0.759962838
#> 70  070 -0.44196244  0.263366158
#> 71  071  1.35330621  1.219261292
#> 72  072 -0.09945554  0.783920146
#> 73  073 -1.40457170  0.026868834
#> 74  074  1.25899586  1.418002986
#> 75  075  0.29641158 -0.683809578
#> 76  076  0.42714320  0.080955817
#> 77  077 -0.58484772 -0.707722658
#> 78  078 -1.31954892 -1.925038260
#> 79  079 -0.07849763 -0.027563273
#> 80  080  0.25431734  1.692263373
#> 81  081 -0.78891454  0.774310169
#> 82  082 -0.53096949 -0.291099155
#> 83  083 -0.52653531 -0.123794145
#> 84  084  0.20261893 -0.294652890
#> 85  085 -0.41382488 -0.592027635
#> 86  086  0.65059773 -0.438474809
#> 87  087 -0.67879732 -2.031451460
#> 88  088 -0.62220900 -0.372695080
#> 89  089 -0.43463816  0.353522713
#> 90  090 -0.25218058 -0.307718791
#> 91  091  0.87564219 -0.269029643
#> 92  092 -1.06719022 -0.327492269
#> 93  093 -0.17619095  0.261188884
#> 94  094 -0.36578494 -1.349347573
#> 95  095 -0.01343477  1.077445966
#> 96  096  0.01780296  0.355594774
#> 97  097 -1.17870570 -1.937333041
#> 98  098 -0.54521076 -0.994406468
#> 99  099  1.34073737  0.156682452
#> 100 100 -1.14030173 -1.752789941