Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4794545
#> Y_2 0.4794545 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.5480690 -0.33198464
#> 2  0.8031929  0.01173450
#> 3  1.7670208  0.06880272
#> 4  0.3564567  0.27505353
#> 5  0.6832778  0.45341309

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.482920079  0.324461376
#> 2   002  1.505408847  0.383470391
#> 3   003  0.145809972 -1.444157971
#> 4   004 -0.968798287 -1.567402181
#> 5   005  0.062995406 -0.940148544
#> 6   006  0.224791095  0.233219521
#> 7   007  0.394644599 -0.575026321
#> 8   008  0.370513507 -1.032807029
#> 9   009 -0.811453709 -0.839366059
#> 10  010 -0.541752490 -0.083007139
#> 11  011  0.436665774  0.431031873
#> 12  012 -0.021987085 -1.052146577
#> 13  013  0.230496889 -0.658632087
#> 14  014 -0.219733763 -0.695294176
#> 15  015 -1.308046735  0.169667416
#> 16  016 -0.437796108 -1.163760726
#> 17  017  0.002035669  0.817371305
#> 18  018 -2.024414293 -1.168604274
#> 19  019  0.472793309  0.243928525
#> 20  020  0.095998489  0.878988373
#> 21  021 -1.811132148 -0.960584851
#> 22  022  0.388008660  0.288936402
#> 23  023  1.408404474  1.771442869
#> 24  024  0.204802689 -0.147428556
#> 25  025 -1.343747845  0.427388075
#> 26  026 -0.647194406  0.593915780
#> 27  027  1.352722885  0.863351907
#> 28  028  0.170061241 -0.576234308
#> 29  029 -1.549668385 -0.789972543
#> 30  030  0.268537636  0.928113361
#> 31  031  0.813097899  0.251379684
#> 32  032  1.310590539 -0.046828854
#> 33  033 -1.561964919 -0.820962851
#> 34  034  0.272342792  0.530131816
#> 35  035 -0.120071184  0.055140338
#> 36  036 -0.096406014 -0.458003945
#> 37  037  0.780514853  1.099761397
#> 38  038 -0.370931665 -0.538549140
#> 39  039  0.151046702 -1.021689231
#> 40  040  0.414009280  1.930241181
#> 41  041  0.318408593  1.299014865
#> 42  042 -0.323141619  0.082819552
#> 43  043  0.246944278 -0.253568802
#> 44  044 -1.313092352 -1.674311462
#> 45  045 -0.377842122  0.430101090
#> 46  046 -0.258659229 -0.464388841
#> 47  047  1.324395829  1.816885808
#> 48  048  1.390363156  0.496195417
#> 49  049 -0.833743602  0.258153931
#> 50  050  0.687438300  2.162533150
#> 51  051 -0.175340904  0.251910045
#> 52  052 -0.008026311 -0.792483660
#> 53  053 -1.947475879 -1.205475110
#> 54  054 -1.286427178 -1.548799330
#> 55  055  0.998225134  0.315735577
#> 56  056 -0.589696598 -0.314195867
#> 57  057  0.291493005 -0.459513892
#> 58  058 -0.768668478  0.096399307
#> 59  059 -0.109983165  1.126535514
#> 60  060  0.127226828  0.223431129
#> 61  061 -0.402826182 -0.848546277
#> 62  062 -1.324486030 -2.702307427
#> 63  063  0.456371296  0.238881565
#> 64  064  1.751299991  0.136668924
#> 65  065 -2.347452657 -0.310293308
#> 66  066 -0.339590648  0.005986809
#> 67  067 -0.984245701 -0.415626308
#> 68  068 -0.649766309 -0.382340137
#> 69  069 -0.270967338 -0.358224675
#> 70  070 -1.341183424  0.593159786
#> 71  071 -0.991532721 -1.591742923
#> 72  072  1.460374320  2.411744887
#> 73  073  0.879121119  0.708263571
#> 74  074 -0.523363159 -1.252065666
#> 75  075  0.454370192  0.660540422
#> 76  076 -1.286136277  0.262920958
#> 77  077 -1.368500323 -0.655725040
#> 78  078  0.078672396 -0.009595237
#> 79  079 -1.663411393 -0.973846910
#> 80  080  2.446857761  1.367681408
#> 81  081  0.152351809  0.780257326
#> 82  082 -0.411297494 -0.798855165
#> 83  083  0.192780914 -1.571990049
#> 84  084 -0.732062152  0.663489902
#> 85  085 -0.120192970 -0.094254225
#> 86  086  0.416683331 -0.382758039
#> 87  087 -0.809351336 -0.740658721
#> 88  088 -2.668352554 -1.934443741
#> 89  089  1.481348689 -1.425166324
#> 90  090 -0.875532448 -1.039372214
#> 91  091  2.114741809  1.615229306
#> 92  092 -1.915778752 -3.106767734
#> 93  093 -0.226874264 -0.700981731
#> 94  094 -0.544806023  0.322901583
#> 95  095  0.142682399  0.589929928
#> 96  096 -1.447674879 -0.396478050
#> 97  097  0.128718955 -1.298455520
#> 98  098 -0.020933832  0.938729943
#> 99  099 -1.682696196 -2.222816533
#> 100 100 -1.316982712 -1.105156905

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.65122589  0.6602639498
#> 2   002  0.80313726 -2.8031861035
#> 3   003 -0.18924795  1.1062896155
#> 4   004  0.23462538  0.7781622455
#> 5   005  0.88780452  0.9920820691
#> 6   006  1.24006305 -1.3394946384
#> 7   007 -0.20072106  0.6499325977
#> 8   008 -1.49627697  0.1213103382
#> 9   009  1.37378045 -0.2957492712
#> 10  010  0.28378495 -0.3677525617
#> 11  011 -0.97628231  0.4187890730
#> 12  012 -0.18897664 -0.4768103414
#> 13  013 -1.20302913 -1.3730148241
#> 14  014 -0.03644498 -0.3108983445
#> 15  015  0.65208478 -0.0536810503
#> 16  016  0.35056764 -0.6487525582
#> 17  017 -0.01764341 -0.4609598884
#> 18  018 -0.30000768  1.4427675983
#> 19  019  1.11175047  0.0055807983
#> 20  020 -1.93874691 -1.7922995266
#> 21  021 -0.89070602 -0.9020572819
#> 22  022  2.42152583  2.3161928808
#> 23  023  0.66965004 -0.5057423331
#> 24  024 -0.12613069 -0.4610699127
#> 25  025  0.18505273  1.5203462046
#> 26  026 -1.08304267 -0.9534912967
#> 27  027 -1.13652682  0.4205377223
#> 28  028  1.34286660  0.0005082841
#> 29  029 -0.13307859  0.1823030791
#> 30  030 -1.46626065 -0.7218498751
#> 31  031  1.25010389  0.7705185837
#> 32  032 -1.03398269 -1.1455741785
#> 33  033  1.82770631  2.1151825348
#> 34  034  1.43548188  0.3984408536
#> 35  035 -0.35080188  1.0920726228
#> 36  036 -1.87796976 -1.7008546919
#> 37  037 -0.91505797 -1.8802040708
#> 38  038 -0.27512284 -1.1528474106
#> 39  039  1.74721903  0.5301371385
#> 40  040  2.46473274 -0.0564751613
#> 41  041  0.91722799  0.7091752637
#> 42  042 -0.63920094 -0.7482744258
#> 43  043  0.56042093  0.7026317754
#> 44  044 -0.46020287 -0.2241661833
#> 45  045  2.27746457  1.7203752424
#> 46  046 -0.06802674 -0.2875952347
#> 47  047  0.86219539  1.3580223289
#> 48  048  0.96310868 -0.5089335064
#> 49  049 -2.52619207  0.7618235657
#> 50  050  0.43437343 -0.0110684512
#> 51  051 -0.07386043 -2.0795155517
#> 52  052 -1.23968948 -0.3083814548
#> 53  053  0.89329984 -0.0629575960
#> 54  054  1.10210358  0.5821475126
#> 55  055  0.90523033  0.4369557355
#> 56  056  0.22368396  0.9439541465
#> 57  057  0.95547581  0.6611088268
#> 58  058 -1.41385626 -0.4292064722
#> 59  059  0.83376994  0.6774296522
#> 60  060  0.92958080 -0.0124241742
#> 61  061 -0.04210236  0.0717623999
#> 62  062 -1.41800339  0.2893154268
#> 63  063 -0.59933755 -0.2184699385
#> 64  064  1.32456851  0.8160984683
#> 65  065 -1.58394580 -0.7834937622
#> 66  066  1.17594395 -0.0602478134
#> 67  067 -0.19055048  0.3658718774
#> 68  068  0.41210142  1.1761866752
#> 69  069 -0.26837724 -0.3681895599
#> 70  070 -0.36790457  1.2992145335
#> 71  071  1.94853198  1.7178005937
#> 72  072  0.42697887  1.7265069642
#> 73  073  1.47198088 -0.1992961454
#> 74  074  0.67484980  0.1631316603
#> 75  075  0.30681399 -0.1167274108
#> 76  076 -1.37928530  0.6882210066
#> 77  077 -0.50227428 -0.5736827019
#> 78  078 -1.17906553 -1.8065555253
#> 79  079  1.62410486 -0.1617037119
#> 80  080  3.29621130  0.4095185097
#> 81  081  0.68682178  1.5900217908
#> 82  082 -0.31521215 -0.3619451471
#> 83  083  0.87967359  0.4882001820
#> 84  084 -2.02150043 -0.7104412489
#> 85  085 -0.17887426 -0.0037695310
#> 86  086  0.18592209  0.7832270131
#> 87  087 -1.15867767  0.0542543765
#> 88  088 -0.17842139  1.1930235257
#> 89  089  2.17381135  0.3236428365
#> 90  090  0.98211429  0.4545956209
#> 91  091  0.11988504  0.6763159108
#> 92  092  0.39500699  0.5059574729
#> 93  093 -1.73858041 -0.4788244784
#> 94  094 -0.73199792 -0.5616501684
#> 95  095 -0.28389688  0.9383894974
#> 96  096 -0.13300052 -0.7168684180
#> 97  097 -0.98543026 -0.8425700640
#> 98  098  0.65194603  0.3482480472
#> 99  099 -1.00802079 -0.8767109091
#> 100 100 -0.10110115 -0.6830475487