Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4655003
#> Y_2 0.4655003 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.02545834 -0.2392571
#> 2 -0.51378899  0.5168598
#> 3 -1.57716888 -0.8449425
#> 4  0.13251071  0.9679881
#> 5 -0.48756018 -1.7454502

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.50814548  0.32926697
#> 2   002 -0.47768692 -1.55052384
#> 3   003  0.72515548  0.47305625
#> 4   004 -1.40151607 -2.19081092
#> 5   005 -1.66888773 -1.37988365
#> 6   006  1.16462804  1.36771962
#> 7   007 -0.38683632 -0.18493339
#> 8   008 -2.26139051 -0.78699767
#> 9   009  1.31370901  1.19278686
#> 10  010 -0.35885384 -0.77576438
#> 11  011  0.92088927  0.31622198
#> 12  012 -0.26141230  0.01784988
#> 13  013  0.45793058  0.99592921
#> 14  014 -0.13030653 -0.22608150
#> 15  015  0.80663642  1.01097762
#> 16  016 -0.35626867 -0.74336925
#> 17  017 -0.63472760 -0.94283960
#> 18  018  0.35415242  0.01095538
#> 19  019  0.28015182  0.03994019
#> 20  020 -0.54671213 -0.14746912
#> 21  021 -0.19329098 -0.06050133
#> 22  022  0.87476509  0.86762295
#> 23  023  0.93749683  0.21496275
#> 24  024 -1.53218154 -0.87256517
#> 25  025  0.97341220  0.93012198
#> 26  026 -0.51657524 -1.66810315
#> 27  027 -0.17144253  1.67589826
#> 28  028 -0.37703634  1.22017313
#> 29  029 -0.17700820  0.62797513
#> 30  030  0.15115713  1.24826519
#> 31  031 -0.82276382  1.31825916
#> 32  032  1.32008019  1.11229304
#> 33  033  0.16572772 -0.76981645
#> 34  034 -0.50183596 -1.24134456
#> 35  035  0.97188542  0.15269673
#> 36  036 -0.47663751 -0.62273640
#> 37  037  0.37650432 -0.20880730
#> 38  038 -0.09349728 -0.63267087
#> 39  039 -0.50825139 -0.88796212
#> 40  040  0.96333591 -0.18901368
#> 41  041  1.42686282  0.83751186
#> 42  042  1.74116100 -0.32739791
#> 43  043  0.47797001  1.01988243
#> 44  044  1.48973766  1.95953520
#> 45  045  1.34649573  0.09043314
#> 46  046 -1.07081154  0.17906279
#> 47  047 -0.76944573 -0.68666418
#> 48  048 -0.48962066 -0.05701960
#> 49  049 -0.47141483 -0.30146401
#> 50  050 -1.05044081 -0.21007815
#> 51  051 -1.15183374 -1.44824706
#> 52  052  0.22664696  0.92941105
#> 53  053  2.59452020  0.49287020
#> 54  054  0.13236444 -0.22780732
#> 55  055 -1.44417516 -1.99722419
#> 56  056 -0.61867587 -0.30089491
#> 57  057  0.35691661 -0.12154195
#> 58  058 -0.13793546 -1.14930569
#> 59  059  1.92559543  0.66118132
#> 60  060  0.30731721  0.22506794
#> 61  061  2.60556723  1.65866833
#> 62  062 -0.35113756 -0.41807119
#> 63  063  1.45375678  0.41276361
#> 64  064  0.96103163 -0.36750379
#> 65  065 -0.34531759  0.33521789
#> 66  066 -1.81248396 -0.70426652
#> 67  067 -0.27814737 -0.88130008
#> 68  068 -1.74943332  0.27431692
#> 69  069 -0.55709684  0.50306257
#> 70  070  0.55671763  0.07472846
#> 71  071 -1.20784423 -0.03452490
#> 72  072 -0.10909225 -0.02956797
#> 73  073  1.00624384  1.35181280
#> 74  074 -0.78464889 -1.37665511
#> 75  075  1.12975893  2.09661974
#> 76  076  0.66879734 -1.50774308
#> 77  077 -0.65572856 -1.34852538
#> 78  078  0.83451036 -0.92168645
#> 79  079  0.94333803 -0.69482272
#> 80  080 -0.71136597 -0.04838313
#> 81  081  1.05795092 -0.48466062
#> 82  082  0.60737775  0.37302231
#> 83  083 -0.13620359 -0.99317772
#> 84  084 -2.66708510 -1.85461661
#> 85  085  0.17093617 -0.63774389
#> 86  086 -1.63061122 -3.25714571
#> 87  087  0.59310118  0.72938117
#> 88  088 -2.34319878 -0.24713794
#> 89  089 -0.60897784 -0.35043492
#> 90  090  0.66419748 -0.16160183
#> 91  091 -1.12560002 -0.48315512
#> 92  092  0.24257482  0.36596169
#> 93  093 -0.77403505  0.10501629
#> 94  094 -0.55086606  0.12569285
#> 95  095  0.01111892 -1.75725892
#> 96  096  0.43736274  0.63986683
#> 97  097  0.18159191  1.96188253
#> 98  098  1.28799688 -0.93497797
#> 99  099 -0.01380668 -0.68565639
#> 100 100 -0.38760811 -0.78886988

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.440264555 -1.102297372
#> 2   002  0.405348138  0.395470556
#> 3   003  0.477554730  0.668882325
#> 4   004 -0.457195355  1.190370644
#> 5   005 -0.663619536 -0.277992643
#> 6   006 -1.005092751 -0.193869164
#> 7   007  0.709589598  0.479269700
#> 8   008  0.809436073  0.825268192
#> 9   009 -1.238402897 -0.616139010
#> 10  010 -0.243321509 -0.032285288
#> 11  011 -1.046479866 -1.696721709
#> 12  012  0.431714984 -0.008167284
#> 13  013 -1.059624646  0.184549581
#> 14  014 -0.285700626 -1.015884890
#> 15  015 -0.956998702 -0.855826940
#> 16  016  0.276248801 -1.376760249
#> 17  017  0.399250357  0.031109289
#> 18  018  0.075182385 -0.063939181
#> 19  019  0.387864894  0.871581593
#> 20  020 -0.653425994  0.303444639
#> 21  021  0.703399818 -0.160018472
#> 22  022 -0.492795973  0.066714133
#> 23  023  1.184352799 -0.063157278
#> 24  024 -0.683725946 -0.257181881
#> 25  025 -0.419872356  0.179974893
#> 26  026  0.147177871  0.395129201
#> 27  027  1.535736235 -0.322024189
#> 28  028  1.265901573  0.722655071
#> 29  029  0.500340241  0.150446080
#> 30  030 -1.340228546 -0.710000195
#> 31  031  1.951593698  0.358439249
#> 32  032  1.182458564  1.370117423
#> 33  033  0.205506261  1.037132172
#> 34  034 -0.521968597 -0.516883831
#> 35  035 -0.227746076  1.072694065
#> 36  036  0.166242969  0.965330948
#> 37  037 -0.203637748 -2.908906659
#> 38  038 -0.624910146 -0.642051510
#> 39  039  0.112636023  0.046259065
#> 40  040  0.847989697 -0.746140073
#> 41  041 -1.108117048 -0.456196280
#> 42  042  0.989316499  1.260120103
#> 43  043 -0.004817894  0.793489182
#> 44  044  0.144888163  0.689425784
#> 45  045  0.072123427 -0.200358078
#> 46  046  0.156691711 -0.184726829
#> 47  047  1.009006097 -0.094587690
#> 48  048  1.473131009  0.850843191
#> 49  049  1.442723200  1.028360979
#> 50  050 -0.691346980  0.085956645
#> 51  051 -0.052150669 -1.055900674
#> 52  052  0.154083801  0.682197488
#> 53  053 -1.043435464  1.007864087
#> 54  054  2.105712426 -0.438628930
#> 55  055  0.869592160  0.654063783
#> 56  056  0.029818977  0.548739550
#> 57  057  0.997432670  1.000566216
#> 58  058  1.275410557  1.647213097
#> 59  059 -0.792859201 -1.159312369
#> 60  060  0.677993629  1.121573266
#> 61  061 -0.568977886 -0.726626174
#> 62  062 -1.370590301 -0.378632376
#> 63  063  0.759680212  0.342476604
#> 64  064  1.804552728  1.410851667
#> 65  065  0.514418269  0.035123010
#> 66  066 -0.525235260 -1.612818750
#> 67  067 -1.070561697 -0.047189889
#> 68  068  0.163456429 -1.295826793
#> 69  069  0.952149011 -0.869975388
#> 70  070 -2.014936603 -1.422797509
#> 71  071 -1.579190680  0.268869763
#> 72  072 -1.798751942 -0.616612271
#> 73  073  0.069275828  2.571157076
#> 74  074  0.258569865  0.193484036
#> 75  075  0.844931417  0.638230115
#> 76  076  1.015292330  0.041273278
#> 77  077  0.314163101 -0.551698482
#> 78  078 -0.081481425 -0.762610864
#> 79  079  1.119693042  0.839101083
#> 80  080 -0.816937681 -0.150452030
#> 81  081 -0.981338376 -0.951864693
#> 82  082  0.617124688  1.796935818
#> 83  083  0.511928197  0.033568975
#> 84  084  0.224705977 -0.281462964
#> 85  085  0.669486053  0.430998149
#> 86  086 -0.408276614  1.174006441
#> 87  087  0.500872941  0.169940775
#> 88  088  0.216707739 -1.242985476
#> 89  089  1.015243517  0.460383631
#> 90  090  1.466789750  2.666373896
#> 91  091  0.597458569  1.630560994
#> 92  092  0.703077938  0.025636655
#> 93  093 -0.606878408 -0.955909328
#> 94  094  1.292935238  0.537750985
#> 95  095 -0.222194206  0.752555460
#> 96  096  0.205570860 -0.246348978
#> 97  097  1.058318051  1.170961957
#> 98  098  1.348343208  0.378773679
#> 99  099  0.484457336 -0.002699214
#> 100 100 -0.652287250  0.197203557