Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.535684
#> Y_2 0.535684 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.1332971  1.04506230
#> 2 -0.4251451 -0.42129175
#> 3  0.6903054  0.02108249
#> 4 -0.1956990  1.49209826
#> 5  0.5226570  1.44447374

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.044192312  0.70946639
#> 2   002  0.494952760  0.18799642
#> 3   003  0.508321008  1.87236371
#> 4   004 -0.237880024  0.38932552
#> 5   005  0.144925712 -1.22652714
#> 6   006 -0.862834438  0.17335231
#> 7   007  0.687160203  0.04308941
#> 8   008 -1.379457372 -0.54726974
#> 9   009 -0.193386958  0.28970571
#> 10  010  0.953097566  0.13524077
#> 11  011 -0.209655627  1.03572328
#> 12  012  1.592667362  0.67680223
#> 13  013  0.077688211 -0.19852859
#> 14  014 -0.177686960  0.97823540
#> 15  015 -1.111229211 -1.00376478
#> 16  016  1.596060221  1.72106463
#> 17  017  0.732058429  0.86001224
#> 18  018 -0.034363835  0.59705623
#> 19  019 -1.475018416 -0.69149423
#> 20  020  1.093742885 -0.59100381
#> 21  021  0.845306639  0.98997991
#> 22  022 -0.556768785  0.47293129
#> 23  023 -0.121200684  0.82288451
#> 24  024 -1.387462784  0.63087456
#> 25  025  0.674142627  1.48701886
#> 26  026 -0.579317469 -0.47707325
#> 27  027  0.768856977  0.93380756
#> 28  028  0.248073048  0.01976684
#> 29  029 -1.644062359 -1.62349632
#> 30  030  1.362378501  1.57488945
#> 31  031  0.628763641  1.31453466
#> 32  032  1.005744411 -1.32133973
#> 33  033  1.279980725  0.20992850
#> 34  034 -0.962886034 -0.79567610
#> 35  035 -1.594851779 -0.94490631
#> 36  036 -0.478184728 -0.02649748
#> 37  037 -0.977598599  0.70657337
#> 38  038  1.717138232 -1.13473623
#> 39  039  0.282707950  0.46578238
#> 40  040  0.305943719  0.78947321
#> 41  041  1.136512920  0.66433378
#> 42  042 -0.656180989  0.73096596
#> 43  043  1.256908570  1.16402356
#> 44  044 -0.606323397  0.70014329
#> 45  045  0.960531619  1.24976917
#> 46  046 -0.006612972  1.78280229
#> 47  047  0.301989402  0.68150464
#> 48  048 -1.180870053 -1.11197834
#> 49  049 -0.274228115 -0.76167357
#> 50  050  0.889884247 -0.98875774
#> 51  051 -0.708655571 -0.58184365
#> 52  052 -1.136091987  0.09879496
#> 53  053 -0.261665937  0.38059997
#> 54  054  1.564379421 -0.58668474
#> 55  055  0.553801898  2.41598057
#> 56  056  0.741392513 -0.51478941
#> 57  057 -0.924616978  0.47269939
#> 58  058 -0.296897807  0.91561721
#> 59  059 -0.459126042  0.83130084
#> 60  060  2.195253612  0.97904527
#> 61  061 -0.348110684  0.09889035
#> 62  062  2.305842357  0.20940932
#> 63  063  1.459366385  0.72999610
#> 64  064  0.618723621  0.65057925
#> 65  065 -0.674963334  1.62140630
#> 66  066 -0.575016331 -0.95298582
#> 67  067  0.594656652 -0.67654576
#> 68  068  0.011753781  1.70016161
#> 69  069 -0.311014254  1.07177364
#> 70  070  0.310002752  0.89799994
#> 71  071 -0.548363197 -0.67059558
#> 72  072  1.104762320  0.44888540
#> 73  073 -0.124264275  1.25319084
#> 74  074 -0.621978937 -0.28396607
#> 75  075  0.923550774  1.09927177
#> 76  076  0.513009220 -1.11940287
#> 77  077 -1.230443198 -0.54222170
#> 78  078  0.342209829  1.68885142
#> 79  079 -2.636589388 -1.18262584
#> 80  080  1.575429294 -0.62361794
#> 81  081  1.293189563  0.53656392
#> 82  082  0.085397031  0.46638822
#> 83  083 -0.171161132  0.38441891
#> 84  084 -1.263309262 -1.39639453
#> 85  085 -1.520825929  0.27595682
#> 86  086 -0.353818655  0.62906995
#> 87  087 -1.998273010 -1.49116169
#> 88  088  0.250484287 -1.22764217
#> 89  089 -0.548159277  1.35792015
#> 90  090  1.592398876  0.62604886
#> 91  091 -2.073402248 -2.11424890
#> 92  092  0.661379524  1.81021281
#> 93  093  0.794606005 -0.77428321
#> 94  094 -0.780683489  0.61986971
#> 95  095 -1.042176700 -0.85808526
#> 96  096  1.139931546 -1.04558372
#> 97  097 -1.156578291 -0.39307964
#> 98  098  1.213069060 -0.37225130
#> 99  099 -0.315780291  0.92875487
#> 100 100 -1.374556718 -0.94321223

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.607863377 -0.07706007
#> 2   002  1.899446988  0.67163239
#> 3   003 -0.790215121 -0.79773434
#> 4   004  0.275709029  0.31922530
#> 5   005  0.437221155  1.25572120
#> 6   006  0.865154054  0.99666366
#> 7   007  2.469121994  2.25738487
#> 8   008 -0.524185219  0.69704732
#> 9   009  0.224450941 -1.58396616
#> 10  010  1.768592081  0.85420687
#> 11  011  0.205503646 -0.82107430
#> 12  012 -0.995401756 -0.58288799
#> 13  013 -0.426839914  0.10168993
#> 14  014 -0.249680539 -1.82769487
#> 15  015 -0.737491461 -0.13562705
#> 16  016  1.896274736  0.73477967
#> 17  017 -0.477030675 -0.65820588
#> 18  018 -2.043044469 -1.37306180
#> 19  019  1.424991472  0.56817283
#> 20  020 -1.095610072 -0.06590173
#> 21  021  0.688917664  0.40331162
#> 22  022 -0.285636365  0.51557479
#> 23  023 -0.333330590  1.18357410
#> 24  024 -0.068456932 -1.08479715
#> 25  025 -1.039335138 -1.74092377
#> 26  026 -2.432270365  0.32669947
#> 27  027  1.841378185  1.08532147
#> 28  028 -0.037986861  0.40750410
#> 29  029  1.451920825  2.20660892
#> 30  030 -1.258387964 -0.81884818
#> 31  031 -0.728068480 -1.51914305
#> 32  032  0.343536122  0.01986591
#> 33  033 -1.828906523 -2.67904454
#> 34  034 -0.176091610 -1.88459001
#> 35  035 -0.934769426  0.28196421
#> 36  036  0.283106781 -0.73868023
#> 37  037 -0.440437833  0.65238453
#> 38  038 -0.096588069 -1.20336210
#> 39  039 -1.147597841 -2.33426906
#> 40  040  1.004129107  0.31931863
#> 41  041  0.005491735 -0.40577498
#> 42  042  0.119314719  1.42334159
#> 43  043 -0.307852643  0.23917902
#> 44  044  0.349305787 -0.04856459
#> 45  045 -1.270876444  0.36502715
#> 46  046  0.661596281  0.04203842
#> 47  047 -0.089321043 -0.53238113
#> 48  048 -0.358002392  0.36368720
#> 49  049 -1.491815534  0.80251112
#> 50  050 -0.665638630 -0.90053068
#> 51  051 -0.427525030  0.55322333
#> 52  052  1.635011366  1.73576677
#> 53  053  0.887750265  0.40434570
#> 54  054  0.700013092  1.15802840
#> 55  055 -0.263789583 -0.38895067
#> 56  056 -0.069248684  0.94874708
#> 57  057  1.104359538  0.54441609
#> 58  058 -0.045826310  0.20336631
#> 59  059 -0.320431297  0.20700714
#> 60  060 -0.665719111  0.23739514
#> 61  061  1.554232651 -0.19115825
#> 62  062  1.571012611 -0.17864776
#> 63  063 -0.922734315  0.28695190
#> 64  064 -0.030663549 -1.04876593
#> 65  065  0.698101828 -1.91472834
#> 66  066 -0.065498233  1.60297938
#> 67  067 -0.360065665  0.09016404
#> 68  068  0.802207078  0.76090938
#> 69  069 -0.632169146 -1.57205569
#> 70  070 -2.390715134 -1.16698152
#> 71  071  1.003379629  0.55620874
#> 72  072 -0.365115073  0.19825692
#> 73  073 -0.698017207 -1.16419540
#> 74  074 -0.162129609  1.80649942
#> 75  075  0.265651147 -0.55225241
#> 76  076 -1.720896630 -1.09165284
#> 77  077 -0.868505688 -1.80844771
#> 78  078  2.449642759  0.64299231
#> 79  079  0.432990488 -1.57549806
#> 80  080  1.604222639  0.77952998
#> 81  081 -0.496566641 -0.34060571
#> 82  082 -0.676400488  0.15067457
#> 83  083  0.751926324  0.47291183
#> 84  084  0.045795404  0.81448672
#> 85  085  0.123535248 -0.35643852
#> 86  086  0.634509292  0.70200490
#> 87  087 -0.081855505  0.36983535
#> 88  088 -0.066877649  0.20154830
#> 89  089 -0.990374727 -0.35810129
#> 90  090 -0.759291070 -0.56063777
#> 91  091 -1.153213858 -1.64917288
#> 92  092  0.626372060  0.85150714
#> 93  093  0.973657616  1.47154575
#> 94  094 -1.681206797 -0.37539955
#> 95  095  1.272481829  0.25139544
#> 96  096 -0.093539528 -0.64923707
#> 97  097  0.446056519  0.21630304
#> 98  098  0.221177945 -1.59418583
#> 99  099  1.088950602  1.40052086
#> 100 100 -0.849374787 -0.77116276