Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5181671
#> Y_2 0.5181671 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.06062651  0.7906761
#> 2  0.07031266 -0.7861782
#> 3 -0.27657364  1.1745773
#> 4 -0.99248183 -0.7974329
#> 5 -0.78161464 -1.5714951

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.970623816 -0.640961915
#> 2   002  1.786715861 -0.601384268
#> 3   003 -1.398025594 -1.372280356
#> 4   004 -0.506190946 -1.335034931
#> 5   005  0.909309099  0.717814760
#> 6   006  1.101025477 -0.298254278
#> 7   007 -0.701358005 -1.442496525
#> 8   008  1.890438492  1.421218499
#> 9   009 -0.317846206  0.212783763
#> 10  010  0.843239703  0.417830551
#> 11  011 -0.065185849  0.379923882
#> 12  012 -1.803034953  0.843842773
#> 13  013  0.619166547  1.989269271
#> 14  014  0.010979819 -0.002535942
#> 15  015 -0.681760696  0.217742396
#> 16  016  0.633422830  0.093847611
#> 17  017  1.404054178  0.204247975
#> 18  018 -0.316074831 -0.665837526
#> 19  019  0.117663198 -0.168436781
#> 20  020 -0.194253320  0.703138008
#> 21  021  0.634521036 -0.640480205
#> 22  022 -1.502324948 -0.256002295
#> 23  023  0.781345235 -0.055606513
#> 24  024 -0.257264525 -0.272573531
#> 25  025  1.509984371  0.035430190
#> 26  026 -0.462033578  0.494488077
#> 27  027 -0.131988652  0.723513926
#> 28  028  0.700525885  0.530397051
#> 29  029 -0.545221229 -0.267393170
#> 30  030  0.369603687 -0.834383879
#> 31  031  0.649613172 -0.846786102
#> 32  032  0.366306253  1.592555374
#> 33  033 -0.367291274 -0.560807255
#> 34  034 -1.065911005 -0.622137701
#> 35  035 -1.192515719 -0.821648104
#> 36  036  0.908032991 -0.081455743
#> 37  037 -1.206446727 -0.702334711
#> 38  038 -1.049451271  0.165539849
#> 39  039  1.728359921  1.022395090
#> 40  040  1.631302383  1.863554129
#> 41  041 -0.521667105 -0.247418384
#> 42  042  1.083012219 -0.155016174
#> 43  043 -0.420281521  1.145039804
#> 44  044 -0.182584374 -1.754408389
#> 45  045 -0.334124207  1.475013115
#> 46  046 -0.676054740 -0.912638165
#> 47  047 -1.489084860 -1.518194697
#> 48  048  0.931067043  1.300372981
#> 49  049  1.240682707 -0.244952369
#> 50  050  1.103505141 -0.771034407
#> 51  051  1.120105898  2.693445729
#> 52  052 -0.937680183  0.178350706
#> 53  053 -0.196004848  1.055682441
#> 54  054 -0.456645190 -0.295559235
#> 55  055  0.966042667  0.735017257
#> 56  056  0.551554635  0.044022294
#> 57  057  0.770904118  0.439115515
#> 58  058  1.315407259 -0.064931056
#> 59  059  0.890460627  0.707706389
#> 60  060 -1.194591508 -1.241570868
#> 61  061 -0.077931726 -0.104672654
#> 62  062 -0.732402968 -0.186810642
#> 63  063 -0.384525061 -1.762864463
#> 64  064 -0.230067066  0.086754126
#> 65  065 -1.322081456 -0.127873669
#> 66  066  1.370566223  1.157346190
#> 67  067  0.808411948  0.785686612
#> 68  068  0.547343906 -1.331689066
#> 69  069  0.331061683 -0.776620930
#> 70  070  0.044843439  0.459835887
#> 71  071 -0.181275968 -0.133713545
#> 72  072  1.182301105  1.469556801
#> 73  073  0.001497512 -0.585041232
#> 74  074 -0.604624536  1.341262464
#> 75  075  0.628628997 -0.498063174
#> 76  076 -1.329528882 -2.550466320
#> 77  077 -0.555903047  0.553704033
#> 78  078  1.109358124  0.896875986
#> 79  079 -0.332843096  0.814932098
#> 80  080  0.196885852  2.003731203
#> 81  081 -0.296643072 -0.449909865
#> 82  082  1.132744712  1.180582687
#> 83  083 -0.069285336  0.509287178
#> 84  084  0.380649533 -1.071404082
#> 85  085  1.072006384 -0.637928550
#> 86  086 -1.026297658  0.533738954
#> 87  087 -0.537201864 -0.650105493
#> 88  088  0.999868118 -1.362467358
#> 89  089  1.061803349 -0.214652984
#> 90  090 -2.095552959 -1.839586152
#> 91  091  0.427368241 -0.551450072
#> 92  092  1.390385963  2.087897785
#> 93  093  1.692773176  0.028926436
#> 94  094  0.769749037 -0.957253753
#> 95  095  2.620258117  1.776206059
#> 96  096 -0.038186976  0.984051250
#> 97  097 -0.728845733  0.066150682
#> 98  098  0.769032215 -0.067781173
#> 99  099 -0.020676238  1.033616485
#> 100 100  2.220565017  1.635780093

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.24834295 -1.464259295
#> 2   002  1.59080383  0.840756249
#> 3   003  0.07102537 -0.577209722
#> 4   004 -0.43258295 -2.152937958
#> 5   005  0.27343820 -1.124982759
#> 6   006  0.38403191  0.029408800
#> 7   007  0.33552516 -0.918022029
#> 8   008  0.46287924  1.146618817
#> 9   009 -0.10517759 -2.316001936
#> 10  010  0.90529325  0.030004803
#> 11  011  0.66919770  2.313700898
#> 12  012 -0.39766512  0.401473269
#> 13  013 -0.20492620 -1.671205716
#> 14  014 -0.78285099 -2.043669193
#> 15  015  0.14399079 -0.446287224
#> 16  016 -0.02879573 -0.087625140
#> 17  017 -1.81774638 -1.789187392
#> 18  018 -0.98762380 -1.255776836
#> 19  019 -0.63322173 -1.572544524
#> 20  020 -0.12880040 -2.758986584
#> 21  021  0.27706985  0.843403349
#> 22  022  1.63786627 -0.805098231
#> 23  023 -1.13604795 -1.113765910
#> 24  024  2.37701517  0.791174192
#> 25  025 -1.85670192 -0.285810495
#> 26  026 -0.21328867  0.480930977
#> 27  027  0.52043434 -1.348058195
#> 28  028 -0.16371201 -1.131539269
#> 29  029  0.04855550  1.098636284
#> 30  030 -0.75676796  0.146314819
#> 31  031 -0.32189629  0.404163129
#> 32  032 -0.44010545  0.012406603
#> 33  033  2.01347875  1.964497542
#> 34  034 -0.43293857  0.423295720
#> 35  035  1.39930018  0.564686914
#> 36  036 -0.58631695 -0.049407800
#> 37  037 -0.45007526 -0.026075151
#> 38  038  0.10774573 -0.078759320
#> 39  039 -0.86004242 -1.239895786
#> 40  040  0.21820086 -0.123344427
#> 41  041 -0.21252376 -0.276468456
#> 42  042  0.01206105 -1.385871576
#> 43  043  0.28678907  0.705268376
#> 44  044 -0.98377181  1.685135651
#> 45  045 -0.66397864 -0.037259613
#> 46  046 -0.09677862 -1.273118139
#> 47  047 -0.65852553  0.039670242
#> 48  048 -0.33819135  0.445307663
#> 49  049 -1.77832189 -1.318702399
#> 50  050 -2.29357711 -0.514637094
#> 51  051  0.21730544 -0.763496012
#> 52  052  0.24829502 -0.477278761
#> 53  053  1.00866289  0.594805222
#> 54  054 -0.13977758 -0.680084007
#> 55  055  0.33952114 -0.059886339
#> 56  056 -0.30490199 -0.415795571
#> 57  057  1.08480041  0.231550756
#> 58  058  0.58017916 -0.969148417
#> 59  059 -0.29121402  0.950241127
#> 60  060  0.36709462 -0.126348541
#> 61  061 -0.59852242  0.380550017
#> 62  062  0.48887613 -0.322956883
#> 63  063 -0.69935060 -0.471861512
#> 64  064  0.29316156 -0.192786914
#> 65  065 -0.36976852 -0.864213110
#> 66  066  1.09432163  0.801790783
#> 67  067 -0.81660264 -1.743694612
#> 68  068 -1.24225770  0.527289599
#> 69  069  0.54733989 -0.979050790
#> 70  070 -0.85197760 -0.152304541
#> 71  071  0.68320192  0.745671560
#> 72  072  0.64642699  1.020162281
#> 73  073  0.62510480  0.046908901
#> 74  074  0.49687539  0.866507503
#> 75  075  0.82285538  1.146209256
#> 76  076  0.34027019  0.271670899
#> 77  077 -0.66512139 -1.005705784
#> 78  078 -0.18014841 -0.115911137
#> 79  079  1.16076191  0.079535589
#> 80  080 -0.22852655  0.085275975
#> 81  081  1.19583909  0.875827491
#> 82  082 -0.50787405 -1.493305687
#> 83  083 -0.03769826 -0.066412349
#> 84  084  0.41675684 -2.173611838
#> 85  085  1.68002328  1.432696804
#> 86  086 -1.22621665  0.425612066
#> 87  087  0.27622977 -0.445104553
#> 88  088 -2.13070131 -1.851693347
#> 89  089  0.17522793 -0.004087839
#> 90  090  0.87740875  0.871627221
#> 91  091  0.36877918  1.276074365
#> 92  092 -0.83027040 -0.679293519
#> 93  093 -0.39347077 -0.079710215
#> 94  094  0.39594433  0.183747708
#> 95  095 -0.35107232 -0.108130844
#> 96  096 -0.65743817  0.891073580
#> 97  097  0.66674633  0.062007424
#> 98  098 -0.10704922  1.300265274
#> 99  099  0.05948951  0.709792107
#> 100 100  0.52676158  1.175110966