Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4892252
#> Y_2 0.4892252 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1        Y_2
#> 1 -1.058496  0.5834790
#> 2 -1.155234 -1.7569298
#> 3  0.517675  0.3484713
#> 4  1.652813 -0.4252194
#> 5  1.250031 -0.3380834

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -2.51830735 -0.697278235
#> 2   002  0.04325679 -0.166789345
#> 3   003  1.19212269  0.787927695
#> 4   004  0.18232227 -0.227682757
#> 5   005 -0.06398171 -1.088125413
#> 6   006 -1.26330894 -1.231359757
#> 7   007  1.28056101  0.823324263
#> 8   008  0.57561268  1.377185259
#> 9   009 -2.15431830 -0.862710444
#> 10  010  0.08838106 -0.919913516
#> 11  011  0.76897121  0.571010498
#> 12  012 -0.29729696 -0.309084057
#> 13  013  0.80919473  0.839700406
#> 14  014 -0.65667173  0.187476949
#> 15  015 -0.57352898  0.422254308
#> 16  016 -0.62835573  1.765260895
#> 17  017  0.03333293  0.337549614
#> 18  018 -2.13289021 -0.169943938
#> 19  019  2.04770754  1.335316837
#> 20  020  0.38890186 -0.868586884
#> 21  021 -1.18676531  0.683459256
#> 22  022 -2.05654808 -2.558588747
#> 23  023 -1.04350150 -1.921789097
#> 24  024  0.09505153 -1.660331719
#> 25  025  0.18775899  1.305425752
#> 26  026  1.35647546  1.537673446
#> 27  027 -1.76075818 -0.628270942
#> 28  028  0.21127921  1.068524492
#> 29  029 -0.99465157  0.110339169
#> 30  030  0.04098195 -0.003378404
#> 31  031  0.02302630 -0.297432397
#> 32  032 -0.43441740 -0.793985338
#> 33  033  2.23364424  1.408737265
#> 34  034  0.47577215  1.806305930
#> 35  035 -2.49991671 -0.629417878
#> 36  036  0.25626422 -0.436737584
#> 37  037  0.24320443 -0.238172134
#> 38  038 -0.33815026  0.841467872
#> 39  039  1.52487814  1.863409910
#> 40  040 -0.94033994 -1.271713177
#> 41  041 -0.09971791 -0.068215891
#> 42  042  1.04779880  1.485785839
#> 43  043  1.30386828  1.237409607
#> 44  044 -0.23896064 -0.280112223
#> 45  045 -0.95076103 -0.237304020
#> 46  046  0.74650247  1.410983371
#> 47  047 -1.24220881  0.101842540
#> 48  048  0.79331487 -0.025617578
#> 49  049 -0.75859297  0.138930717
#> 50  050  0.94462835  0.105113523
#> 51  051  1.56587199 -0.317943736
#> 52  052 -0.81580302 -0.492300094
#> 53  053 -1.23475725 -0.589878066
#> 54  054 -1.00156954  0.319102823
#> 55  055 -0.99808115  0.258910423
#> 56  056 -0.22464932  0.300810703
#> 57  057  0.31755946  1.571505534
#> 58  058  0.19176603 -0.022115908
#> 59  059  0.15172422 -0.388525094
#> 60  060  0.12941933  1.259927689
#> 61  061 -0.25596465 -1.162274382
#> 62  062 -1.17212468  0.112385930
#> 63  063 -0.61760972  0.757837191
#> 64  064 -0.01345981  1.161064619
#> 65  065 -1.73841008 -2.140588223
#> 66  066 -1.18154611 -0.073273041
#> 67  067  0.05541403 -0.031821431
#> 68  068 -2.21621032 -1.484139382
#> 69  069 -0.55086609 -0.845673365
#> 70  070 -0.44506945 -1.054094854
#> 71  071  0.84037313  1.112665360
#> 72  072 -0.21440235 -0.356899726
#> 73  073 -0.47146736 -1.556820829
#> 74  074  1.70190487  1.278894163
#> 75  075  1.23438091  0.643820959
#> 76  076 -0.99186425 -0.579877473
#> 77  077  0.08854792 -0.033515565
#> 78  078 -0.12413758  0.238400246
#> 79  079  1.15315537  0.309902672
#> 80  080  0.77485530  0.474517748
#> 81  081  1.49877178  0.057405101
#> 82  082 -2.16465845 -1.148259863
#> 83  083 -1.68967487 -1.052126020
#> 84  084  2.15949759  0.615930217
#> 85  085  0.96094902 -0.185281385
#> 86  086 -0.52507344  0.794769134
#> 87  087 -1.10582611 -0.138901390
#> 88  088  1.15228567 -0.914637965
#> 89  089 -1.07473792 -0.726384310
#> 90  090  1.19286724  0.394216613
#> 91  091 -0.24204407  0.052084846
#> 92  092 -1.35186585 -1.535076365
#> 93  093  0.68363731  0.462309122
#> 94  094 -0.74423065  0.880087979
#> 95  095  1.01943008  0.126970231
#> 96  096  0.16089744  1.096222322
#> 97  097  2.25182278  1.529085746
#> 98  098 -0.31539146 -0.715924789
#> 99  099 -1.23683478 -1.404355402
#> 100 100 -0.57115129 -0.735103478

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.03263494  1.17423904
#> 2   002 -1.17932665 -0.92613164
#> 3   003  2.25615575  0.04580734
#> 4   004 -0.63556864  0.15637763
#> 5   005  0.82231302 -0.77558904
#> 6   006  1.12696263  1.16813266
#> 7   007 -1.17621043 -0.16621704
#> 8   008 -0.46211398 -0.35129810
#> 9   009 -0.69867632 -1.02711325
#> 10  010  1.72473669 -0.06442113
#> 11  011 -1.49553223 -1.72719030
#> 12  012  1.13341860  0.26195027
#> 13  013  2.13710012  1.89301258
#> 14  014 -0.24440022 -0.28096150
#> 15  015 -0.54715399  0.84237838
#> 16  016  1.36206663  1.10646823
#> 17  017 -0.30648826 -1.92805208
#> 18  018 -1.73990688 -0.99508156
#> 19  019  1.07614621 -0.72956153
#> 20  020 -0.20266335 -1.74135154
#> 21  021  1.02839697  0.79188678
#> 22  022  0.28047190 -0.62415898
#> 23  023 -0.24593466 -1.18907357
#> 24  024 -0.01227788  1.01664650
#> 25  025  1.29021934 -0.17289967
#> 26  026  1.44462495  0.10353671
#> 27  027  1.22812625  1.78822624
#> 28  028  0.81896254 -0.34780710
#> 29  029  0.21280960 -0.45563662
#> 30  030 -1.00716152 -1.53686237
#> 31  031 -0.02709802  0.58398019
#> 32  032  1.36871343  2.22944408
#> 33  033  0.79913119 -0.44739388
#> 34  034 -0.85005293 -0.90138889
#> 35  035  0.73471741  1.03109263
#> 36  036 -1.85201900 -1.44003159
#> 37  037 -0.02576792  1.35221949
#> 38  038 -0.41740084 -0.77426740
#> 39  039 -0.21427068 -0.32748134
#> 40  040 -0.14866601 -0.13423394
#> 41  041  0.08977085  0.41916139
#> 42  042  0.50440006 -1.79265539
#> 43  043 -2.28893114 -1.71792907
#> 44  044  0.52684239  0.46323526
#> 45  045  0.63777896  0.41868921
#> 46  046  1.51074108 -0.06697984
#> 47  047  0.77173488  0.70375611
#> 48  048  1.67335967  0.20264834
#> 49  049 -0.19639663  0.17791561
#> 50  050  0.23223354  0.08667309
#> 51  051  1.17545162  0.65930023
#> 52  052  0.70986107  0.96008937
#> 53  053 -0.03488354  0.65776253
#> 54  054 -0.47870106 -0.97560215
#> 55  055  0.32666440 -0.12395256
#> 56  056  2.44033850  0.99924243
#> 57  057  0.37805830 -0.43425727
#> 58  058  2.05661169  1.35018520
#> 59  059 -0.52267038 -1.55152487
#> 60  060 -0.70649913  0.34907597
#> 61  061  1.26515504  0.09020568
#> 62  062  0.89559776 -1.29327591
#> 63  063 -0.23454152  0.20896166
#> 64  064 -0.64396698 -1.46087954
#> 65  065  1.64789364  0.15182095
#> 66  066 -0.15450529 -1.02020248
#> 67  067 -0.89304322 -0.76570195
#> 68  068  2.22631676  0.89614549
#> 69  069  1.24515210  0.15563160
#> 70  070 -0.27180065  0.60299747
#> 71  071 -1.11106260 -0.52419308
#> 72  072  1.58201426  1.53721578
#> 73  073 -0.22144789 -0.79931876
#> 74  074  1.05632266  0.32471817
#> 75  075 -1.14941367  0.27439437
#> 76  076  0.23661080 -0.11957533
#> 77  077  0.32958160  1.03934027
#> 78  078  0.72516597  0.65762369
#> 79  079  2.70558752 -0.05827960
#> 80  080  1.32031970 -0.61197921
#> 81  081  0.11657528  0.28830240
#> 82  082 -0.40821220 -0.25118228
#> 83  083 -0.49375953  0.24181017
#> 84  084  0.12082980 -0.51349488
#> 85  085 -0.38493161  0.31714951
#> 86  086  0.22221294  1.63596637
#> 87  087  0.72672378 -0.85608747
#> 88  088 -0.27205086 -0.16063155
#> 89  089 -0.24208749 -0.57436705
#> 90  090 -3.02581246 -0.65504446
#> 91  091  1.59938836 -0.18557806
#> 92  092 -0.58922021  1.68696589
#> 93  093 -0.10527332  0.11203910
#> 94  094 -0.80939874 -0.41811123
#> 95  095 -1.10249657  0.35995855
#> 96  096  0.75072729  0.74433137
#> 97  097 -0.34115409  1.06248974
#> 98  098  0.13717274 -0.63267145
#> 99  099  0.60838077 -0.87324901
#> 100 100  0.05587512  1.00799842