Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4870471
#> Y_2 0.4870471 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.4484451 -1.1555831
#> 2  1.5864576  2.8020026
#> 3  1.2818290 -0.7397735
#> 4  0.9362471  1.2734119
#> 5  1.6208520  2.0241526

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1          Y_2
#> 1   001  2.7140957410  3.234056806
#> 2   002  0.6595170823 -0.435544107
#> 3   003  2.1684405097  0.830822452
#> 4   004  0.7451103618  0.717746208
#> 5   005  0.6609592791  0.295053596
#> 6   006  0.4478108891  1.524531362
#> 7   007 -0.4405734932 -0.005803729
#> 8   008  0.8280464411 -1.112269849
#> 9   009  0.6001565378 -0.744071434
#> 10  010  0.5652078509  0.090537983
#> 11  011  1.4610089240  0.919275770
#> 12  012 -1.1748369003 -0.786867982
#> 13  013 -0.8623679073 -1.049693857
#> 14  014 -0.2006740295 -2.483329911
#> 15  015  0.5521417858 -0.530554544
#> 16  016 -0.2993724496 -0.778452844
#> 17  017  1.1998680192  2.374592363
#> 18  018  0.6643886911 -1.877530310
#> 19  019 -1.8117206340 -1.251482828
#> 20  020 -0.3886114371  0.423988247
#> 21  021 -0.5911979548  0.051194356
#> 22  022  0.7691454802  1.237643124
#> 23  023  1.1386521454  0.724252181
#> 24  024 -0.3797627904  1.000600346
#> 25  025 -0.3687095481 -0.562585737
#> 26  026  0.9600281347  1.757775566
#> 27  027  2.8221545602  1.784635816
#> 28  028 -0.6780921095 -1.669383828
#> 29  029 -0.6739039648 -1.668932740
#> 30  030 -0.1821710045 -0.313281737
#> 31  031 -0.7219658873 -1.258685971
#> 32  032  0.5456547804  1.035501268
#> 33  033 -0.0372193114  1.365521450
#> 34  034 -0.8998668785 -0.063090846
#> 35  035  0.4318949456  1.472190141
#> 36  036 -0.1091025188 -1.846531248
#> 37  037 -1.4563910136 -0.490000171
#> 38  038  1.0235166748  0.902452544
#> 39  039  0.7848324219 -1.343178435
#> 40  040  2.2933726735 -0.405698561
#> 41  041 -1.2955450439  0.656147152
#> 42  042  0.3115068473 -1.860935888
#> 43  043 -1.1062808217  0.844088100
#> 44  044 -1.1433012329 -1.215713526
#> 45  045 -0.0595111534  0.322272932
#> 46  046 -0.3154277866 -0.115353574
#> 47  047  0.6279118572  1.015858163
#> 48  048 -0.1331999294  0.964260703
#> 49  049 -0.2899714771  0.045838257
#> 50  050 -0.6056208994 -0.255047749
#> 51  051  1.1820206043  1.655527183
#> 52  052 -0.0344869260 -0.418812505
#> 53  053 -1.2956254752 -1.820427402
#> 54  054 -0.8599801591 -1.011833855
#> 55  055  1.7812670096  1.004701056
#> 56  056  1.2139354677  0.686540627
#> 57  057 -0.7975019493 -0.683162700
#> 58  058 -0.5235460436  1.318219225
#> 59  059  0.0000654703 -1.191015790
#> 60  060 -1.0257111970  1.043244984
#> 61  061  0.0400924650 -0.376049860
#> 62  062  1.1106577302  0.149832330
#> 63  063  0.0756818524  0.567024091
#> 64  064 -0.3501331425 -0.217806540
#> 65  065 -0.0057573464  0.060998599
#> 66  066  0.1799338226  1.067353412
#> 67  067  0.7606567674  2.241529966
#> 68  068  1.8021273993  1.429091001
#> 69  069 -2.3453742015 -1.637795778
#> 70  070  0.0887025058 -1.095804072
#> 71  071 -0.7773843908 -0.579879021
#> 72  072  0.1903360827  1.984971862
#> 73  073  0.4246024016  1.039674411
#> 74  074 -0.2230774557 -1.950522601
#> 75  075  0.3945307130  0.481156885
#> 76  076 -1.0816277602 -0.784428864
#> 77  077  0.0548136676 -0.112990950
#> 78  078  0.0890642236  0.613091646
#> 79  079  0.7032248478  1.363551577
#> 80  080 -0.1637083773 -0.927945749
#> 81  081  0.9681447378  0.053713791
#> 82  082 -0.2178660505  0.105119808
#> 83  083  0.1378419741 -0.380670163
#> 84  084  0.3618478779  0.052450020
#> 85  085  1.4553404777  0.070719517
#> 86  086  0.3686500804 -0.589195725
#> 87  087 -1.3013178397 -0.961563309
#> 88  088 -1.0129122798 -1.469075634
#> 89  089  2.1538033370  2.326466025
#> 90  090 -0.1291398466  0.411384321
#> 91  091 -0.7974047975  0.770063802
#> 92  092 -0.2107199415 -0.285266586
#> 93  093 -1.0191840877 -0.930637239
#> 94  094 -2.1332210652 -0.801102469
#> 95  095  0.6049357348  0.584749938
#> 96  096  0.0656352776  1.776275692
#> 97  097  0.8959153418  0.780153662
#> 98  098 -0.1505786506  0.828530917
#> 99  099  0.4090818868 -0.044997972
#> 100 100  0.1957638050  0.255195314

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.083970298 -0.35270141
#> 2   002  0.115734231  0.05519948
#> 3   003 -0.938667151 -0.75688850
#> 4   004 -1.015689423 -0.49697446
#> 5   005 -0.225723576  0.59641734
#> 6   006 -1.324565096 -0.96136840
#> 7   007  2.799710797 -0.10264988
#> 8   008  0.741131253  0.05720513
#> 9   009  1.121393382  0.73777252
#> 10  010  0.412016791  0.78003390
#> 11  011  0.784346578  0.07582159
#> 12  012 -1.076735820  0.12384321
#> 13  013 -1.844645253 -0.63222471
#> 14  014 -0.593832815 -2.20270625
#> 15  015  0.915705719  1.46094743
#> 16  016  1.229783235  1.03793813
#> 17  017 -1.250525539 -1.69099235
#> 18  018  0.723203968  2.58188578
#> 19  019  1.119712549 -1.05329833
#> 20  020 -0.817074593 -0.20226330
#> 21  021  0.846348744 -0.02661575
#> 22  022  0.458368419  0.23096514
#> 23  023  1.555602204  0.37147510
#> 24  024 -1.424076017 -0.78312079
#> 25  025  0.289204922 -1.89828647
#> 26  026 -1.085217732 -2.15649954
#> 27  027 -0.043571542 -0.09877382
#> 28  028  0.078631819  0.43076161
#> 29  029  0.376676958 -1.18675136
#> 30  030 -0.025269043  0.26648093
#> 31  031  0.424639133 -0.69768553
#> 32  032  0.379393357  2.23242220
#> 33  033 -1.182141233 -1.11796958
#> 34  034 -0.612161421  1.17314596
#> 35  035 -2.250010323 -0.82326820
#> 36  036  1.883620775  0.33274416
#> 37  037 -0.634017778 -1.23180323
#> 38  038 -0.058203407  1.53843149
#> 39  039  1.561431045  0.19365684
#> 40  040 -1.381755810  0.78005583
#> 41  041  0.621289084  2.10499027
#> 42  042 -0.248047289 -0.16099566
#> 43  043 -1.184123404 -1.60108625
#> 44  044  1.036933013  1.08761367
#> 45  045  1.689473280  1.38959207
#> 46  046  0.681014671 -0.92657697
#> 47  047  0.597626115 -1.01480660
#> 48  048 -0.322037059 -0.63203234
#> 49  049 -0.033518062 -0.04667599
#> 50  050 -0.846557530 -1.15814925
#> 51  051 -1.355460174 -0.98855527
#> 52  052 -1.409858643 -1.21967207
#> 53  053 -0.998139615 -0.75246820
#> 54  054 -1.167339835 -1.96590318
#> 55  055  0.628806974  0.19395521
#> 56  056  0.429883347  0.93505067
#> 57  057 -2.325451316 -1.10415256
#> 58  058 -0.394662910  0.87440464
#> 59  059  1.439163545  0.27896595
#> 60  060 -1.591726299 -0.36249697
#> 61  061 -0.169829888 -0.21699891
#> 62  062 -1.391778732 -0.52875905
#> 63  063  1.145389219  1.38270901
#> 64  064  0.924977801  0.43519999
#> 65  065 -0.030882076  1.33448797
#> 66  066  1.051015182  0.20581711
#> 67  067 -1.441568732 -0.11430988
#> 68  068  0.207518035  0.55773753
#> 69  069  0.729058173  0.57979044
#> 70  070 -1.285603898  0.30136186
#> 71  071  0.462341118  1.43560656
#> 72  072 -1.146572285 -1.80325788
#> 73  073 -0.626662092 -0.37026211
#> 74  074  0.406370826  1.26038323
#> 75  075 -1.545631735 -0.55496436
#> 76  076  0.615837801 -0.37117480
#> 77  077  0.929143763  0.65009303
#> 78  078  1.885952308 -1.01396526
#> 79  079  1.438147949  1.72094482
#> 80  080 -0.093670265 -0.09577291
#> 81  081 -0.092168081  0.18628408
#> 82  082 -0.546246965 -0.26284539
#> 83  083 -0.820232816 -0.59000670
#> 84  084 -1.300685363 -1.06013693
#> 85  085 -0.730665352 -0.11251363
#> 86  086 -0.961903619 -1.37397448
#> 87  087  1.242267041  0.77234360
#> 88  088  0.826331335 -0.74115900
#> 89  089  0.347732881 -0.81530865
#> 90  090  1.304549299  0.23542588
#> 91  091  0.938895857  1.18586077
#> 92  092 -1.309832245 -0.15208185
#> 93  093 -0.002330534  1.13626842
#> 94  094 -0.061755156 -1.56191657
#> 95  095 -0.693889311  0.33967803
#> 96  096  1.331105992  0.96420045
#> 97  097 -0.104302280  0.46854578
#> 98  098 -0.325614934  0.55496408
#> 99  099  0.780208467 -0.27624199
#> 100 100  0.277527441  2.05046225