Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4688129
#> Y_2 0.4688129 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.05574301  0.4060537
#> 2 -0.70933639 -0.3083302
#> 3  0.39386162  0.4225352
#> 4 -0.51618394 -2.4711705
#> 5 -0.13143631 -0.9100432

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.13121920 -0.129825197
#> 2   002  0.05479739 -1.405642444
#> 3   003 -0.12791826 -1.919526051
#> 4   004  0.84803714  0.246938806
#> 5   005  1.61580645  0.459476291
#> 6   006  1.17546200  0.778125665
#> 7   007 -1.02747117 -1.267610619
#> 8   008 -1.13086523 -0.502196464
#> 9   009  0.36620854  0.026938000
#> 10  010 -0.74147498 -0.006385554
#> 11  011 -1.04867164 -0.728722069
#> 12  012 -1.24134490 -1.796248193
#> 13  013  1.62778828 -0.345561071
#> 14  014  1.75268549  0.528450137
#> 15  015  1.99619835  1.577525051
#> 16  016  0.93557326  0.276335320
#> 17  017  0.47410853  1.358332121
#> 18  018  1.61594012  0.133399528
#> 19  019  2.76211952  1.898577242
#> 20  020  0.14943311  0.121996341
#> 21  021  0.38134570  0.618666413
#> 22  022  1.42587128  2.124864225
#> 23  023 -1.48729303 -0.276230529
#> 24  024 -0.24543888  0.602054475
#> 25  025  0.08663063  0.161042685
#> 26  026 -0.07681573 -0.276736562
#> 27  027  0.63242592 -1.246463909
#> 28  028 -0.82231481  0.559293836
#> 29  029  0.56885594 -0.745028619
#> 30  030 -0.75270764  0.407520008
#> 31  031  0.56011649 -1.241318459
#> 32  032 -0.26261551 -0.342981239
#> 33  033  0.20250918  0.353252621
#> 34  034  1.57799628  1.161304559
#> 35  035  0.05409302 -0.097885062
#> 36  036  0.87949384  2.710370663
#> 37  037  1.58451634  1.562773889
#> 38  038  1.95615526 -0.659930050
#> 39  039 -0.12044066  0.816704333
#> 40  040  1.57654090 -0.096281838
#> 41  041  0.80724965 -0.080817564
#> 42  042  0.25057161 -0.025920342
#> 43  043  1.86536751  0.773179397
#> 44  044  1.06748930 -0.361664761
#> 45  045  0.42433854  1.863529540
#> 46  046 -1.31282370 -0.212886343
#> 47  047  1.04216161  1.524954574
#> 48  048  1.37703733  0.671615530
#> 49  049 -0.05709373 -1.733281242
#> 50  050  0.76041645 -0.651445101
#> 51  051 -0.22309821 -0.385230880
#> 52  052  1.22533621  1.515665715
#> 53  053 -0.57141893  0.022561712
#> 54  054 -0.28241373  1.479693986
#> 55  055  1.69008034  0.903294213
#> 56  056 -0.20570901 -0.318148349
#> 57  057  0.89953942 -1.368090866
#> 58  058 -0.07846520 -1.266552650
#> 59  059 -0.11707725  0.548682450
#> 60  060 -0.59942351 -0.113914093
#> 61  061  0.11742296  0.264250246
#> 62  062 -0.90680035  0.002849171
#> 63  063  0.44341126 -0.269290942
#> 64  064  0.29722729 -1.999276190
#> 65  065  0.12129748  0.520924046
#> 66  066  0.02198871 -0.586556781
#> 67  067  0.25476205  0.129658843
#> 68  068 -0.63523134 -1.727989569
#> 69  069 -0.12560779  0.715853207
#> 70  070  1.57023932 -1.154059079
#> 71  071  1.15651801  1.315624874
#> 72  072  0.67545523  0.880317636
#> 73  073 -0.09953974 -2.008941161
#> 74  074 -0.19803441 -0.992772391
#> 75  075  2.42977015  2.631284608
#> 76  076 -0.88384371 -0.584671441
#> 77  077 -0.43212439 -0.703617926
#> 78  078  1.98688307  0.390529591
#> 79  079 -0.79584297 -0.169844262
#> 80  080  0.97061077 -0.178773546
#> 81  081  0.27276306 -0.029101918
#> 82  082  1.00337736  0.076393178
#> 83  083  0.60907348  1.484334237
#> 84  084 -0.05597476  0.171914343
#> 85  085  0.21152661 -0.437785358
#> 86  086  0.45108236  1.433501392
#> 87  087  0.69313306  0.876059062
#> 88  088  1.73726575  1.687776489
#> 89  089  0.38806976 -0.660649075
#> 90  090 -0.33826612  0.011316582
#> 91  091 -0.19458137 -1.329332965
#> 92  092 -0.09626678 -1.044376955
#> 93  093 -0.63196628 -0.176748873
#> 94  094  0.24027822  1.063222712
#> 95  095  0.20616430  1.408092703
#> 96  096  1.19129858  1.029851872
#> 97  097 -0.60128494 -0.768742766
#> 98  098 -0.13258335  1.874355473
#> 99  099  0.25561561  0.236699334
#> 100 100 -0.34805012 -0.270505731

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1          Y.2
#> 1   001 -0.6494662736 -1.103671216
#> 2   002  0.9705739239  1.043610514
#> 3   003 -2.8931798526 -0.927498072
#> 4   004  0.7008880030 -0.058597074
#> 5   005 -0.9411725731  0.538604511
#> 6   006  0.1687687267 -0.182715668
#> 7   007  0.9817808363  1.096285979
#> 8   008 -0.5895043412 -0.508063489
#> 9   009  0.6236811843 -0.849497929
#> 10  010  0.7637359106  0.065097973
#> 11  011 -0.4723277138 -1.105969189
#> 12  012  0.9853037092  1.012473003
#> 13  013  1.4826783029  2.356474158
#> 14  014  0.4128693140  0.524783478
#> 15  015  1.6492470116  0.381442543
#> 16  016 -0.9203678391 -0.633058538
#> 17  017  1.0580521814 -0.154844398
#> 18  018  0.2191050350  0.725393107
#> 19  019 -2.3856876937 -0.918098379
#> 20  020 -0.9140634417 -1.759874220
#> 21  021  0.5597895183  0.929690229
#> 22  022  1.5655597504  0.565607790
#> 23  023  1.1580841010  0.179723042
#> 24  024  0.3308195311  1.167054702
#> 25  025  0.6890304427  0.545038371
#> 26  026 -0.0357346419  0.918960308
#> 27  027  1.1812211922  1.331857248
#> 28  028  0.7041266398 -0.353076950
#> 29  029  0.6473675620 -0.901719462
#> 30  030  0.0354741505  0.072334197
#> 31  031  1.5280304233 -1.897039054
#> 32  032  0.8298642751  0.463607348
#> 33  033  0.5442588956  0.769495036
#> 34  034 -0.6345587784  0.229638465
#> 35  035  0.9052157876 -0.410105738
#> 36  036 -0.0003577348 -0.750620598
#> 37  037 -0.7226616225 -1.228261167
#> 38  038  1.0984095961  0.254422042
#> 39  039  1.3990011271  1.659441575
#> 40  040 -0.3239331306 -0.937535668
#> 41  041  0.8184096462  1.190016939
#> 42  042  0.5652160729  1.233466634
#> 43  043  1.2380882935  0.050142738
#> 44  044 -0.1183735554  0.370429342
#> 45  045  0.1233461828  0.921935300
#> 46  046  2.0146231034  1.452724713
#> 47  047 -0.7517573677 -1.074050345
#> 48  048 -0.3721839025 -0.880896384
#> 49  049  0.0073867046  0.005057776
#> 50  050  0.7987139345  1.833966018
#> 51  051 -2.7081103910 -2.488382263
#> 52  052 -1.1009374399 -0.030896963
#> 53  053 -2.0384890111 -1.183280534
#> 54  054 -0.0610664415 -1.278843963
#> 55  055 -0.8451817900 -0.162222443
#> 56  056 -0.7775379089  0.165861569
#> 57  057 -0.5043066800 -1.017220621
#> 58  058 -0.4917472986 -0.716520900
#> 59  059  1.8449816140  1.664293907
#> 60  060  0.5821693988  0.272679162
#> 61  061 -1.6514173214 -1.466205408
#> 62  062  0.0401536443  1.179399796
#> 63  063  0.3020328560  0.873543683
#> 64  064 -0.5641855922 -1.332411112
#> 65  065 -0.6464134223 -0.547841430
#> 66  066  1.4604750795  1.162007609
#> 67  067 -0.9393078473 -0.387885994
#> 68  068 -0.9143333747 -0.315136960
#> 69  069  0.6266305201  1.645483779
#> 70  070 -0.0753671690 -0.866030753
#> 71  071 -1.4105843175 -0.906440578
#> 72  072 -0.3807656972 -1.395753686
#> 73  073 -0.3705847802 -0.338722844
#> 74  074  1.6956911895  0.599293466
#> 75  075 -0.4784354000  0.461006203
#> 76  076  2.1387583396  1.795877339
#> 77  077  1.1448404036  1.513108716
#> 78  078  1.0789589696  0.961455127
#> 79  079 -0.7156906683 -1.308801328
#> 80  080 -0.2276775465  0.354666373
#> 81  081  0.4821323459 -1.485574094
#> 82  082 -1.8149583138 -1.332623537
#> 83  083  0.0723438303  0.710749172
#> 84  084 -0.4720588046 -0.109052742
#> 85  085  0.6798345554  1.158611423
#> 86  086  0.5200810518 -0.710401973
#> 87  087 -0.1761565399  0.724753924
#> 88  088 -0.1698109825 -0.792753089
#> 89  089  1.0335779825  0.561621256
#> 90  090  0.0606422075  0.570903851
#> 91  091  0.1798679681 -0.675073669
#> 92  092 -0.1971191094 -0.708920767
#> 93  093 -0.4357099846 -1.419918336
#> 94  094 -1.4252986457 -0.030261632
#> 95  095  1.6644678492  3.012676309
#> 96  096  0.2037080591 -1.437709094
#> 97  097 -0.5195196102 -0.384082586
#> 98  098  0.3104106482 -0.595232111
#> 99  099 -0.0905099989 -1.186331336
#> 100 100  0.5716192067  0.450090198