Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.517462
#> Y_2 0.517462 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3077026 -0.5704862
#> 2 -0.8780030 -1.0853542
#> 3  1.6058865 -0.3663463
#> 4  0.3163092 -0.4576447
#> 5  2.2545352  0.7805723

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1         Y_2
#> 1   001  0.9115170871  1.40581556
#> 2   002 -0.3968113108 -0.31282321
#> 3   003 -1.2914343916 -0.92239606
#> 4   004 -0.6490358317  1.11330890
#> 5   005  0.1285513439  0.56817357
#> 6   006 -1.5271673197  0.32109329
#> 7   007  0.4523034032 -0.98641053
#> 8   008  0.4682553136  0.27667871
#> 9   009  0.4594506028 -0.36941576
#> 10  010  1.0564256270  1.39740087
#> 11  011  1.8255882327  1.25996833
#> 12  012 -0.1513578671 -0.35796264
#> 13  013 -0.1820302849 -0.56583093
#> 14  014  2.3682976884  0.91252930
#> 15  015  0.4817118776 -0.28665937
#> 16  016  1.3089539210  0.34794986
#> 17  017 -0.6700422613 -0.93048659
#> 18  018  0.2642687856 -0.77080368
#> 19  019 -1.9573946530 -2.04850257
#> 20  020 -0.4699242715  0.30028954
#> 21  021  1.9594383285  1.18180438
#> 22  022 -0.8096456425 -0.27327253
#> 23  023  0.0904528323 -0.15740137
#> 24  024 -2.1616629351 -0.74927813
#> 25  025  0.4715037789 -0.02586433
#> 26  026 -0.7933735154 -1.44789666
#> 27  027 -0.8926264408  1.19740056
#> 28  028  1.8716782500  1.53325982
#> 29  029 -0.9842547478  0.15874029
#> 30  030  0.4438324961 -2.26363563
#> 31  031  2.0825152470  0.50285093
#> 32  032 -0.5460774803 -0.71129947
#> 33  033  1.0658071673  0.06209896
#> 34  034  1.9712441591 -0.01723876
#> 35  035 -0.9399061193 -1.61596118
#> 36  036 -0.6810495606 -0.16634835
#> 37  037  0.1109486346 -0.58042511
#> 38  038  0.8246527941  1.99844745
#> 39  039 -1.6948987633 -1.52876903
#> 40  040 -0.3811006078  0.57964595
#> 41  041  0.4105548132  2.02952791
#> 42  042 -0.9469103684 -1.57254556
#> 43  043 -1.1960038991 -1.69150185
#> 44  044 -0.2411644365  0.13512425
#> 45  045  1.8951834719  1.31026373
#> 46  046  0.6411788490 -0.02105576
#> 47  047 -0.2611224616 -1.46736215
#> 48  048  0.5355457673  0.26047348
#> 49  049  0.1148723717  0.91234319
#> 50  050 -0.1211884767 -0.06244159
#> 51  051  0.6011977284  0.98907477
#> 52  052  0.6099379884  0.49424703
#> 53  053  0.4508278744  0.24998991
#> 54  054  0.7574916776  1.09827778
#> 55  055  0.4566459623  0.09514066
#> 56  056  0.9276447128  1.14522212
#> 57  057  1.4140793483  0.92657506
#> 58  058 -0.0024653593 -0.52624529
#> 59  059  1.1399744944 -0.39470216
#> 60  060 -0.3961824997 -0.50439043
#> 61  061  0.0004145897  0.09306486
#> 62  062  0.9826187513 -0.84516443
#> 63  063 -0.3390156734 -1.27902715
#> 64  064 -0.6869392735 -0.43631467
#> 65  065  0.0743165446  0.32276381
#> 66  066 -0.8950302479 -0.05752296
#> 67  067  1.4093531382  1.41755138
#> 68  068 -1.2713295025 -2.25188733
#> 69  069  0.3299022218 -1.06895241
#> 70  070 -1.1623813031 -1.38103719
#> 71  071  0.4113437891 -1.04506138
#> 72  072  0.4459230320 -0.04232221
#> 73  073 -0.0603025704  0.15270229
#> 74  074  0.7547271065 -1.06606911
#> 75  075 -1.2378537017 -0.12177804
#> 76  076  0.7433535578  1.45986406
#> 77  077  0.7516464108  0.53604554
#> 78  078  0.8892501154  1.09382938
#> 79  079 -0.0311907650 -0.19978592
#> 80  080 -0.7991258251 -0.37104874
#> 81  081 -1.1498583865 -1.46030366
#> 82  082  0.0158770022 -1.03918766
#> 83  083  0.6342766034  0.40310858
#> 84  084  0.0958029561  1.37388537
#> 85  085 -0.3030165326  2.17948473
#> 86  086 -0.1723840367  0.79191211
#> 87  087 -0.7396674121  1.56076928
#> 88  088 -2.3330955234 -2.95786327
#> 89  089  0.2522410755  0.29600670
#> 90  090 -0.1572352392 -0.02219610
#> 91  091  1.0618971507  1.16216781
#> 92  092  1.1269164664  0.93129508
#> 93  093  0.3674617434  0.61174227
#> 94  094 -0.7940957739 -1.38106309
#> 95  095  0.0323666050  0.27351741
#> 96  096  0.0999307269  1.55086125
#> 97  097 -0.0201139060 -1.58751805
#> 98  098  0.9782294094  0.45976774
#> 99  099 -0.3797398611 -1.06792306
#> 100 100 -1.2364110762 -0.40558766

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.994523184  2.13346897
#> 2   002 -2.807570386 -1.00838361
#> 3   003 -1.094115053 -1.47280866
#> 4   004  0.191796834  0.21706015
#> 5   005  0.998793545  0.75086115
#> 6   006 -1.153975788 -0.32955099
#> 7   007  0.188397118  0.16595834
#> 8   008  1.668815470  2.37935739
#> 9   009  0.187723579 -0.30312126
#> 10  010 -1.853522602 -0.82867197
#> 11  011 -3.257387480 -1.72821677
#> 12  012 -0.226441197  1.17895357
#> 13  013 -0.147305215  0.46379011
#> 14  014 -0.929678736 -0.04014215
#> 15  015  0.168868887 -1.87430381
#> 16  016  0.603359080 -0.79359698
#> 17  017  1.338125905  0.11967977
#> 18  018 -1.072032597 -0.65893523
#> 19  019  1.312785941  0.46027335
#> 20  020 -0.232494141  0.14248560
#> 21  021  1.984666481  2.36542223
#> 22  022  1.165611966  1.91786621
#> 23  023 -0.831678216 -1.09665473
#> 24  024  1.243641977 -0.53696695
#> 25  025  0.367972256 -0.94125084
#> 26  026  0.977644966  0.68440138
#> 27  027 -0.169525231 -1.03572434
#> 28  028  0.210587938  1.59614234
#> 29  029  1.210542210  0.66560585
#> 30  030  0.316168254  0.46873612
#> 31  031  0.491654252 -0.58922116
#> 32  032 -1.808970341 -0.95606226
#> 33  033 -0.566479145 -0.42416733
#> 34  034  0.494383815  2.06245744
#> 35  035  0.987975470 -0.56815458
#> 36  036  2.264576913  2.37589401
#> 37  037  0.094436632  0.92058275
#> 38  038 -1.504103791 -1.91096030
#> 39  039  2.128629105  1.22937914
#> 40  040  0.808688543 -0.43269124
#> 41  041  0.346617016  1.09204919
#> 42  042 -0.766989532 -2.35229464
#> 43  043 -2.043036422 -1.00706573
#> 44  044  0.205291070 -0.49958230
#> 45  045  0.835407248  0.27675925
#> 46  046 -1.675974894 -0.92756161
#> 47  047 -0.243338252  0.13673397
#> 48  048  0.743453411 -0.06095228
#> 49  049  0.797154858  0.56778094
#> 50  050 -0.371291298 -1.37845656
#> 51  051 -1.431017074 -0.05063421
#> 52  052  1.097820672  1.98815818
#> 53  053 -0.292383672 -0.48327209
#> 54  054  1.753921257  0.17543997
#> 55  055  0.718781572 -0.40573947
#> 56  056  0.985239929  0.76857845
#> 57  057 -0.006240641  0.80254866
#> 58  058  0.456949939  0.52755658
#> 59  059 -0.064504512  0.27897047
#> 60  060  0.778461944  0.20088827
#> 61  061 -3.044589210 -1.26558822
#> 62  062 -0.830396875 -0.30671056
#> 63  063 -0.727406467 -1.77184440
#> 64  064 -0.115363327  0.02991041
#> 65  065  0.736178987 -0.35191617
#> 66  066  0.985399874  0.21257269
#> 67  067 -1.328180201  0.11540364
#> 68  068 -1.041871642 -0.77039216
#> 69  069  0.089913208  0.72448858
#> 70  070 -0.666088544  1.47373238
#> 71  071  0.994141871  2.01222564
#> 72  072 -0.896050627 -1.32148808
#> 73  073 -1.942706795 -1.03127096
#> 74  074 -1.010495129 -0.84209379
#> 75  075 -0.479900385 -0.93128202
#> 76  076 -0.017682427 -0.42062655
#> 77  077  1.593646089  1.59768716
#> 78  078 -1.870059538 -0.76664641
#> 79  079  0.675785431  0.68385683
#> 80  080  0.003827194  0.50838029
#> 81  081 -0.628675405 -0.59369099
#> 82  082  2.236216412  1.16130817
#> 83  083 -2.378755534 -1.72685626
#> 84  084  0.021709729 -0.41233640
#> 85  085 -0.306754218  0.09373545
#> 86  086  0.708589119  1.45405032
#> 87  087  0.426897822 -1.45325936
#> 88  088  0.557655912 -1.62416229
#> 89  089  0.097370391  1.70727744
#> 90  090  0.123607337  0.29539890
#> 91  091  1.657199511  0.78717266
#> 92  092 -1.543029489 -2.45780785
#> 93  093  0.827338169  1.44825593
#> 94  094  1.067980117  0.34012579
#> 95  095 -0.534975205  1.27113225
#> 96  096 -0.540785082  0.23592854
#> 97  097  1.351257634  0.71894407
#> 98  098 -0.602435788 -0.91141470
#> 99  099 -0.001614692  0.61040847
#> 100 100 -1.002326318  0.47573298