Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4801231
#> Y_2 0.4801231 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3905475  0.1326519
#> 2  0.6483072 -1.0095261
#> 3 -0.2751298 -1.3004036
#> 4 -0.3462939  1.4943054
#> 5  2.1874169  1.3396332

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.961150053 -0.50048529
#> 2   002  0.102828841  0.46413574
#> 3   003  1.145237670  0.80199985
#> 4   004  0.708002539  0.74610857
#> 5   005  1.666112689  0.46858834
#> 6   006  1.072797149  0.61533147
#> 7   007  0.276197331  0.59593908
#> 8   008  0.320861208 -1.40425185
#> 9   009  0.144173661  0.34511818
#> 10  010  0.035368229 -1.03855860
#> 11  011 -0.524581220 -0.05325765
#> 12  012  1.735581173  0.19971925
#> 13  013 -1.019613765  0.10855900
#> 14  014 -0.095881452  1.22187277
#> 15  015  0.654688332 -0.18495855
#> 16  016  1.226630987  0.57396324
#> 17  017  0.926475073 -0.19722354
#> 18  018 -0.102882991 -0.53379366
#> 19  019 -0.303370424  1.11561847
#> 20  020 -1.133298541 -1.79256403
#> 21  021 -1.867045982 -0.96805311
#> 22  022 -0.323998079 -1.16960149
#> 23  023  0.949355006  0.99863582
#> 24  024  0.447883032  0.64224547
#> 25  025 -0.259731304 -0.15632498
#> 26  026  1.740135388 -0.36584039
#> 27  027  1.818188364  0.73581406
#> 28  028  0.978227619  0.96075447
#> 29  029 -0.532261169 -0.97375958
#> 30  030 -0.127450655  0.01859338
#> 31  031 -1.289693327 -0.64046231
#> 32  032 -0.643227396 -1.14548825
#> 33  033 -1.563769171 -0.08734616
#> 34  034 -0.001879005  1.21583104
#> 35  035  0.007299244  0.55062779
#> 36  036  1.071995319 -0.40429230
#> 37  037  2.723795307  1.12473900
#> 38  038  1.901964944  0.91298709
#> 39  039  0.857981077  0.78969226
#> 40  040  0.256641314  0.42485638
#> 41  041  0.263883672 -0.28565025
#> 42  042  1.664586798  1.55535355
#> 43  043  0.638265598 -1.02410282
#> 44  044 -0.835329056 -0.46129563
#> 45  045 -1.694753629 -1.36799511
#> 46  046 -1.077094756 -1.57399116
#> 47  047  0.778899363  1.10200068
#> 48  048 -0.760916218 -0.25834740
#> 49  049 -0.091296319 -0.64208163
#> 50  050 -0.293951023  0.46405356
#> 51  051  0.150581533  0.25911228
#> 52  052  1.012466819 -0.80641852
#> 53  053  0.425621177 -0.25046119
#> 54  054  0.897725958 -0.05551869
#> 55  055 -0.857901508 -0.39631780
#> 56  056 -0.313757039 -0.16986750
#> 57  057 -2.889768186 -1.09389647
#> 58  058 -0.117476011 -2.09520192
#> 59  059  0.624395143  0.71704115
#> 60  060 -1.006185380 -0.12840251
#> 61  061  0.034528768  1.50890813
#> 62  062  0.650230457  0.77551438
#> 63  063  0.455675172  0.80108729
#> 64  064  1.864483884  0.38117934
#> 65  065  0.624294981  0.06820107
#> 66  066 -1.591085104 -1.10671239
#> 67  067  0.236215682  0.85626905
#> 68  068 -0.019327143 -0.06009409
#> 69  069 -2.095524930 -1.44296450
#> 70  070  0.753406126 -1.14032842
#> 71  071  0.719047003 -1.40488872
#> 72  072 -0.049233439 -0.16137446
#> 73  073  0.242944608 -0.64091601
#> 74  074  1.220532645  1.21412332
#> 75  075 -0.442259614 -0.43637730
#> 76  076  2.176261457 -0.51750326
#> 77  077 -0.207002957  1.28654277
#> 78  078  0.534834065  0.04399175
#> 79  079  0.316245520  1.31816264
#> 80  080  1.020654580 -1.05644246
#> 81  081 -0.579480525 -0.58528534
#> 82  082  0.021131726  0.73726904
#> 83  083 -0.239305758  0.19125387
#> 84  084 -0.844194561 -0.38881534
#> 85  085 -1.253270197  0.02011754
#> 86  086 -0.578490385 -0.27901013
#> 87  087  0.362209708 -1.35041134
#> 88  088 -1.632722432 -1.93582678
#> 89  089 -1.057572589 -1.33653639
#> 90  090 -0.153661278 -1.08663971
#> 91  091 -0.698229644  1.26037654
#> 92  092 -0.322274381  1.08020053
#> 93  093  0.543755407  1.26927254
#> 94  094  0.126165458  0.16945919
#> 95  095 -0.024679037 -0.87986969
#> 96  096  1.030138930  0.60560177
#> 97  097 -0.926424346 -1.39175730
#> 98  098 -0.066482790  1.13873764
#> 99  099  0.448639367 -0.80658392
#> 100 100 -0.693199109 -1.91079792

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.41257253  1.375780770
#> 2   002 -0.75220838 -1.521127148
#> 3   003  0.67819925  1.989931887
#> 4   004 -1.07338812  0.653969578
#> 5   005  0.39125867  0.489333945
#> 6   006 -0.31250155  0.799858617
#> 7   007 -0.85819880  0.225801528
#> 8   008  0.75641096  2.016652985
#> 9   009 -0.85828385  0.155840243
#> 10  010 -1.02384004 -1.332862732
#> 11  011  0.52208014  1.012605032
#> 12  012 -1.34656267 -1.734325347
#> 13  013  1.01998372  1.253889689
#> 14  014  1.49188199  0.914939907
#> 15  015  0.32450030 -0.156671736
#> 16  016  1.65757405  2.656328440
#> 17  017 -0.88855800 -1.003271406
#> 18  018 -1.62094423  0.673910930
#> 19  019 -0.99678782 -1.107012523
#> 20  020 -0.27756575 -0.124581795
#> 21  021 -0.72519162 -0.553852325
#> 22  022  0.81369223  1.091983700
#> 23  023 -0.16164645  0.878691522
#> 24  024  1.00564949  0.354494309
#> 25  025  1.40473478  0.333187493
#> 26  026  0.06636684  0.813803093
#> 27  027 -1.55863466  0.837173422
#> 28  028 -0.25722887  0.642063253
#> 29  029 -0.30225821 -0.769135452
#> 30  030  0.38529386 -0.870064278
#> 31  031  0.93569602 -0.715413989
#> 32  032 -0.14340926 -1.472009374
#> 33  033 -0.67418261  0.312446068
#> 34  034  0.82046911  1.064944521
#> 35  035  0.43333792 -0.444803443
#> 36  036  0.53518312  0.276668488
#> 37  037  0.08511807 -1.740514563
#> 38  038 -0.37204870 -0.226446260
#> 39  039  0.80087732  0.175426416
#> 40  040  0.11786409 -0.079525983
#> 41  041  0.38070231  1.131623863
#> 42  042 -0.07188744  0.723670318
#> 43  043  0.34794703  1.094771056
#> 44  044 -1.32731365  0.200457034
#> 45  045 -0.59429123  1.304239953
#> 46  046 -0.05043660  1.078632841
#> 47  047  1.37530387  0.876810927
#> 48  048 -1.16828842 -0.707469543
#> 49  049 -1.38663462 -0.109303622
#> 50  050  1.50023702 -0.288144856
#> 51  051 -0.37241888  0.348522574
#> 52  052 -0.82431977 -0.284965303
#> 53  053  0.23480661  0.024044101
#> 54  054  0.76561618 -0.167188982
#> 55  055 -0.27178936 -0.321808922
#> 56  056 -0.55540322  0.493372492
#> 57  057  0.11034509 -0.360089975
#> 58  058 -1.33861498  0.383074953
#> 59  059 -0.27924756  0.951723613
#> 60  060 -0.81238799  0.872624662
#> 61  061  0.73424276  1.113517023
#> 62  062  0.61315969  0.152817005
#> 63  063  0.56197058  0.166076858
#> 64  064 -0.07677734 -0.110451704
#> 65  065  0.37526457  3.364035518
#> 66  066  1.29258531  0.404855547
#> 67  067  2.67319852  2.160395639
#> 68  068 -0.99986991 -0.700325584
#> 69  069  0.71340653 -0.336454532
#> 70  070 -2.06432674 -0.756744931
#> 71  071  0.54540492  1.310969378
#> 72  072  0.13940933  0.714983508
#> 73  073  0.89221066  1.944490598
#> 74  074 -1.77377130  0.267697569
#> 75  075  0.07741158  0.211086798
#> 76  076  0.40796204  1.338749580
#> 77  077 -0.54885885 -0.493179243
#> 78  078  0.75883091 -0.528665064
#> 79  079 -0.87337391 -0.293682435
#> 80  080  1.56560906  0.638279105
#> 81  081 -0.64154818 -0.974606501
#> 82  082 -0.06132280 -1.234167950
#> 83  083  0.35573285 -0.776308794
#> 84  084  1.73816671  1.214184659
#> 85  085 -1.08189674  0.997405947
#> 86  086  0.19545134 -0.699738557
#> 87  087  0.91627311  0.699981384
#> 88  088  0.90401098  0.594501886
#> 89  089  0.01318547 -1.286404496
#> 90  090  0.08359402  0.370552327
#> 91  091 -1.80194239 -0.001541375
#> 92  092 -0.11141141  0.429399613
#> 93  093 -0.43521276 -1.815519979
#> 94  094 -0.35852314 -0.098232753
#> 95  095 -0.27571954  0.965890834
#> 96  096  2.85311128  3.360517217
#> 97  097  0.19026048 -1.243069235
#> 98  098  1.16737783  0.951369369
#> 99  099 -1.62922916 -0.248340072
#> 100 100  0.55611915  1.709818287