Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.508938
#> Y_2 0.508938 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1       Y_2
#> 1  1.1252225 0.4240544
#> 2 -1.0838671 0.4518383
#> 3 -0.2480781 0.1601694
#> 4  2.4814353 1.4788207
#> 5  0.9443337 1.6014463

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.807342767  1.60643821
#> 2   002  0.259333894  0.75343970
#> 3   003 -0.819593788  0.55461265
#> 4   004  0.948196134  1.18412194
#> 5   005 -1.709784045 -0.07652385
#> 6   006  0.330689787  1.73561259
#> 7   007  0.427032140  0.78933990
#> 8   008  1.001848616 -1.14056049
#> 9   009 -0.341735468  0.05571819
#> 10  010  1.187563663  1.87396757
#> 11  011 -0.494087191  0.33450973
#> 12  012  0.449684723  1.48099160
#> 13  013 -0.106694947 -0.74700976
#> 14  014  0.996731688  0.48952325
#> 15  015 -0.485987878 -1.37081473
#> 16  016  1.187506467  1.73939531
#> 17  017  0.316707764  0.27778215
#> 18  018  0.054681392 -0.99172313
#> 19  019 -0.973739374 -0.63354833
#> 20  020  1.316737579  0.58594453
#> 21  021 -0.118528241  0.18226774
#> 22  022  0.390047370 -0.79977369
#> 23  023  0.987964059  0.38162923
#> 24  024  0.793057985  1.29422854
#> 25  025 -1.383443226 -1.91679191
#> 26  026  0.171806990  1.12474167
#> 27  027 -0.851962235 -1.01408870
#> 28  028 -1.122938822 -0.21100440
#> 29  029  1.821220860  1.88785501
#> 30  030 -0.151838931 -0.07505640
#> 31  031 -0.423035465 -0.95984085
#> 32  032 -0.509764033 -0.46113694
#> 33  033 -0.253088093  0.01932455
#> 34  034  1.484152843  1.01625648
#> 35  035 -0.902438382 -0.92679590
#> 36  036 -0.637828440 -0.51979511
#> 37  037 -0.532114923 -1.29799814
#> 38  038 -0.005279967 -0.13585338
#> 39  039  0.522966602 -0.01208657
#> 40  040  0.582327742  0.25855391
#> 41  041  0.107540490 -0.91321341
#> 42  042  1.302111498  0.43494462
#> 43  043 -1.857494139 -0.33273994
#> 44  044 -0.655657530 -1.73190713
#> 45  045 -0.289220264  0.10970486
#> 46  046 -0.130628149  0.63650598
#> 47  047  0.851756721  1.94088053
#> 48  048  0.867745841 -0.52280696
#> 49  049  0.398609288 -0.01877585
#> 50  050 -0.918694013  0.61897832
#> 51  051  0.859066634  0.91834986
#> 52  052 -1.376604473 -0.24485064
#> 53  053 -0.524134884 -1.70573779
#> 54  054  0.820878974 -0.38483814
#> 55  055 -0.643929877 -0.77464263
#> 56  056  0.339503720  0.52500627
#> 57  057  0.136054674 -1.17289085
#> 58  058 -0.290672052  2.30621759
#> 59  059  3.043761228  1.56740889
#> 60  060 -0.759818975  0.30934390
#> 61  061 -1.244731424  0.49549570
#> 62  062 -0.008672123  0.25955795
#> 63  063  1.736878272  1.35366863
#> 64  064  0.339011561  1.12022693
#> 65  065 -1.322878832 -0.27143992
#> 66  066 -1.302551563  0.04364674
#> 67  067  0.778779838 -0.63475843
#> 68  068  0.200968252 -0.38121827
#> 69  069  1.613721585  0.55143277
#> 70  070 -1.728741324 -0.28981690
#> 71  071  1.738837196  2.04220937
#> 72  072 -1.623243988 -0.49031009
#> 73  073 -0.730055303 -0.69110726
#> 74  074  0.225717922 -0.04250105
#> 75  075 -1.093937545  0.19981650
#> 76  076  1.166534514  1.89989277
#> 77  077  0.459188046  0.49872536
#> 78  078  0.531093116  1.42139890
#> 79  079  0.325185248  1.00979673
#> 80  080 -0.516305233 -0.38138826
#> 81  081  0.717519824  0.50811479
#> 82  082 -0.002893458  0.62451120
#> 83  083 -0.847357088 -0.69636018
#> 84  084  0.230017136 -0.47266682
#> 85  085 -0.762868951 -0.45994450
#> 86  086  1.256618416  0.79628008
#> 87  087  1.517654405  0.10611734
#> 88  088  0.301455650  0.84525342
#> 89  089  1.125733375  0.03572313
#> 90  090 -0.041602853  1.01926062
#> 91  091 -1.320556907 -0.74272941
#> 92  092  1.161280175  0.17887755
#> 93  093  0.608394258  1.06546662
#> 94  094 -0.327077257 -2.15332115
#> 95  095 -0.903571992  0.48216381
#> 96  096  0.533491134  0.96049425
#> 97  097 -0.093434380  1.78649997
#> 98  098 -0.388813320 -1.33903022
#> 99  099  0.465658750 -0.32948063
#> 100 100 -0.211635415  0.69611350

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.72460796  1.03325127
#> 2   002 -1.31889931 -1.31836006
#> 3   003  0.07917064  2.01549396
#> 4   004 -0.75502478  0.50570162
#> 5   005  0.84957577  0.43644582
#> 6   006 -2.21159071 -0.22056939
#> 7   007 -1.04379745 -1.86656306
#> 8   008  0.34294441  1.44757090
#> 9   009 -0.93475865 -0.55084503
#> 10  010  1.41372864 -0.99790240
#> 11  011 -1.12735051 -2.35871605
#> 12  012 -0.63309644  0.24814691
#> 13  013 -1.18033238 -0.28545608
#> 14  014  1.43079497  0.40776288
#> 15  015 -0.74277667  0.09473117
#> 16  016 -0.83851486  0.31521155
#> 17  017  1.22859924  1.02813757
#> 18  018  2.22483948  1.21991651
#> 19  019 -0.44508977 -0.95782540
#> 20  020 -0.61811655  0.03475429
#> 21  021 -0.81365578 -0.55692907
#> 22  022  0.02769730 -0.06086661
#> 23  023  2.05283571  1.83680685
#> 24  024  1.36339628 -0.29460726
#> 25  025  2.63604340  0.40470668
#> 26  026 -0.14184062  0.40173952
#> 27  027 -0.72258520  0.10720292
#> 28  028 -1.26533161 -0.02019895
#> 29  029 -0.63934298  0.20981514
#> 30  030 -0.41526752 -2.36693005
#> 31  031  1.14503395  0.50645704
#> 32  032 -0.96449490 -0.84551991
#> 33  033 -0.37322830  1.88345333
#> 34  034  0.71577204  1.15465341
#> 35  035  1.92042733  1.70986673
#> 36  036  0.17215297  0.69513570
#> 37  037 -0.09572266  1.87013387
#> 38  038 -0.30855268  0.43752751
#> 39  039  0.38997961 -0.70940513
#> 40  040  0.02001153  1.45658580
#> 41  041 -0.94455009 -2.12455935
#> 42  042 -0.73319604 -0.05914395
#> 43  043 -1.24147463 -0.21015843
#> 44  044  0.32727199  1.32077835
#> 45  045  0.42622272  0.63483150
#> 46  046  0.63391579  0.07380902
#> 47  047  0.77935296 -0.66586587
#> 48  048  0.17054693  0.95878523
#> 49  049  1.66671136  0.41651544
#> 50  050  0.09572037 -0.51869444
#> 51  051 -0.69434802 -0.37716613
#> 52  052 -0.48006801 -0.51115437
#> 53  053  0.27248924 -0.44914419
#> 54  054  0.65972023  0.05059575
#> 55  055 -1.29307047 -2.41685692
#> 56  056  0.02542262  0.12547278
#> 57  057 -0.59723004 -1.81175493
#> 58  058  0.15142156  2.10788262
#> 59  059  0.43480077 -0.07924000
#> 60  060 -0.81486923 -1.70154767
#> 61  061  0.93059219  0.70331737
#> 62  062  0.26952962  1.74404904
#> 63  063 -0.06012836  1.69928337
#> 64  064  0.43704075  2.21426283
#> 65  065 -1.04175212 -1.43059365
#> 66  066 -0.64660750 -0.17924939
#> 67  067 -0.02718947 -1.16568101
#> 68  068  1.44743820  0.69723683
#> 69  069 -0.58674911 -1.00318501
#> 70  070 -1.35580755  0.06420435
#> 71  071 -0.98161991 -1.66817992
#> 72  072 -0.96251340 -0.29732749
#> 73  073  0.14565546  0.72557548
#> 74  074 -1.07984234 -1.46715688
#> 75  075  0.18758020  0.59055459
#> 76  076  1.33302012  1.83527928
#> 77  077  0.54364607  0.71108900
#> 78  078  0.71298293 -0.13827414
#> 79  079 -0.95736593  0.89462957
#> 80  080 -1.41605004 -1.03626934
#> 81  081 -0.26347361 -0.95984253
#> 82  082 -1.67874673 -0.88402021
#> 83  083  0.61443103 -0.91359671
#> 84  084 -0.30309919  0.94972213
#> 85  085 -1.77347885 -0.26907667
#> 86  086 -0.65094881 -0.62789805
#> 87  087  0.35966542  0.13894799
#> 88  088  0.05789451  0.03268248
#> 89  089  0.13640294  1.63692636
#> 90  090 -0.34070068 -0.86363025
#> 91  091 -1.80064344 -1.48229933
#> 92  092  0.82254453  0.30705773
#> 93  093  0.30281973  0.75709398
#> 94  094 -0.39793312  0.48962286
#> 95  095  0.24131009  0.06112407
#> 96  096  0.09055476  1.57238433
#> 97  097  0.10826272 -1.67186295
#> 98  098  0.78599818 -0.11408461
#> 99  099 -1.11382451 -1.74658199
#> 100 100  0.85537812 -1.57380968