Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5126198
#> Y_2 0.5126198 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.7328111  1.4565020
#> 2  2.0148476  1.0114392
#> 3  0.9838455  0.3921842
#> 4  0.7448529 -0.0359544
#> 5 -1.1082595 -0.2981294

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.66457657  1.487064210
#> 2   002 -0.49571161 -2.357859679
#> 3   003  1.14917052  1.187500555
#> 4   004  0.95620485  1.057213877
#> 5   005  0.28122694 -0.008320835
#> 6   006  0.94398928  0.516392930
#> 7   007  0.59168795  0.772463109
#> 8   008  0.54312514  0.837280919
#> 9   009  0.58532143  1.592911837
#> 10  010 -0.48718235  0.868328728
#> 11  011 -1.10090236 -0.110057667
#> 12  012  0.69909186  0.205182611
#> 13  013  2.18513553  1.319518845
#> 14  014  0.16765358 -0.484961653
#> 15  015 -0.29739232 -0.490476485
#> 16  016 -0.74594315  0.907969138
#> 17  017  1.32251038  0.341286659
#> 18  018  0.57729057  0.718087411
#> 19  019  1.53665086 -0.624537503
#> 20  020 -2.10508561 -0.610628705
#> 21  021 -0.59479955 -0.534732762
#> 22  022 -2.33213930 -1.367462970
#> 23  023  0.10039849 -0.287931117
#> 24  024 -0.24130023 -0.027534915
#> 25  025  0.41281461  1.859890472
#> 26  026  0.25972831 -0.442053732
#> 27  027  0.59621652  0.268425688
#> 28  028 -0.38979199 -0.017620444
#> 29  029  0.04370982  0.614214446
#> 30  030 -0.68860923 -2.243188615
#> 31  031 -0.34143128  0.369297052
#> 32  032 -0.77325188 -0.392177918
#> 33  033  0.78927587 -0.242262781
#> 34  034  0.18963473  0.883323674
#> 35  035  0.58654673 -1.211810824
#> 36  036 -0.20263052  0.647787527
#> 37  037  1.11299422  1.632077247
#> 38  038  2.46944177  1.780609937
#> 39  039  1.57878921  1.537430865
#> 40  040  1.05992227  0.336903791
#> 41  041  2.16622395  1.293551211
#> 42  042 -0.69349235 -0.207000312
#> 43  043 -0.04495510 -0.861814292
#> 44  044 -0.67422421 -0.474805460
#> 45  045  2.01530733  0.480646728
#> 46  046 -1.83549345 -1.027723240
#> 47  047  0.20385798 -0.770374871
#> 48  048  0.24468776  0.107630862
#> 49  049  1.21881063 -0.556011856
#> 50  050  0.84512089  0.044850980
#> 51  051  0.80445930  0.024867845
#> 52  052  0.75206786  0.107405495
#> 53  053  1.73819811 -0.132628905
#> 54  054  0.42991194 -0.027393507
#> 55  055  0.93608740  1.334195250
#> 56  056 -0.06317119  0.736479509
#> 57  057  1.50949948  1.603296625
#> 58  058  1.07806729  1.934769423
#> 59  059 -0.30373840 -0.357044743
#> 60  060 -1.78521416 -0.640375390
#> 61  061  0.22308896  0.011680654
#> 62  062  0.04843037 -0.974244267
#> 63  063  0.46107725  0.797397952
#> 64  064  0.93893042 -0.272999468
#> 65  065  0.17459858 -0.401721788
#> 66  066  0.40568289 -1.119056683
#> 67  067 -1.29560320  0.183210196
#> 68  068  0.24184269 -0.826930712
#> 69  069 -0.17151583  0.305271073
#> 70  070 -0.21114419  0.874792796
#> 71  071 -1.33032808 -1.122592514
#> 72  072 -1.74632542 -1.458333350
#> 73  073 -1.35205369 -1.537382964
#> 74  074 -1.36897575  0.570112056
#> 75  075 -1.02486904 -1.489651784
#> 76  076  0.37541077  0.641960257
#> 77  077 -0.38493527 -0.556443736
#> 78  078  0.14628423 -1.014149484
#> 79  079 -1.42549694 -1.607143173
#> 80  080  0.18626971  0.133450479
#> 81  081 -0.94538276 -1.566435379
#> 82  082 -1.01502833  0.283723745
#> 83  083  0.08788483  0.740282519
#> 84  084 -1.51356326 -1.228433896
#> 85  085 -1.49552462 -1.774970105
#> 86  086  0.12066578 -0.382474101
#> 87  087 -1.40497178 -0.521091368
#> 88  088 -0.02777860 -0.137504177
#> 89  089  1.10509295 -1.850330457
#> 90  090  0.93320399  0.052954515
#> 91  091 -0.66796975 -0.915404686
#> 92  092 -0.38477612  1.409137874
#> 93  093  0.45577986  1.047533361
#> 94  094  0.05631651  0.087462543
#> 95  095 -0.56204835 -0.260821267
#> 96  096 -0.48591941  0.594274389
#> 97  097  0.74884928 -1.725415932
#> 98  098  0.57357228  0.808934040
#> 99  099  0.07347781  1.091993442
#> 100 100  0.87440079  0.195415709

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.74604733  0.42680383
#> 2   002  3.22097484  1.44137209
#> 3   003 -1.90884720 -0.41673897
#> 4   004 -1.43197746  1.56812672
#> 5   005 -0.06765577  0.33810136
#> 6   006 -0.01675211  0.58138677
#> 7   007 -0.98608305 -0.88570988
#> 8   008  0.09898895 -0.17916929
#> 9   009  0.24152361 -1.05693104
#> 10  010  1.66688413 -0.37592131
#> 11  011  0.99243414  0.74153247
#> 12  012  0.63299049 -0.89302690
#> 13  013 -0.43765929 -0.11245865
#> 14  014 -0.36270273  0.02865888
#> 15  015  1.74397120  1.42956074
#> 16  016 -0.33757850  0.81251938
#> 17  017  2.79713867  1.69813147
#> 18  018 -0.89915141 -0.92101284
#> 19  019  0.18721231  1.19914301
#> 20  020  0.54755578  0.09420854
#> 21  021  2.17035249  0.69665712
#> 22  022 -0.27358804 -0.82549146
#> 23  023 -1.18719796 -0.45564144
#> 24  024 -0.01724755  0.27330497
#> 25  025 -1.16689199 -0.81994889
#> 26  026 -0.87709574 -0.21457190
#> 27  027  0.13638012  0.11867840
#> 28  028 -0.25480091 -1.46404414
#> 29  029  0.76995360 -0.15622758
#> 30  030  0.29604853  0.24213577
#> 31  031  0.27509749  1.20144475
#> 32  032 -1.80688685 -0.51579183
#> 33  033 -0.12821563 -0.15502260
#> 34  034  1.59268548  0.14388950
#> 35  035  0.46719385 -0.04605497
#> 36  036  1.06367989  2.58971783
#> 37  037  0.33320648 -1.24608776
#> 38  038  1.78714788  1.01227611
#> 39  039 -0.50586497  0.29213414
#> 40  040  0.22094513 -0.43853330
#> 41  041 -1.91327618 -0.86426620
#> 42  042 -1.76303693 -0.22498450
#> 43  043 -0.68225721  0.76307170
#> 44  044  0.16366467  0.63391073
#> 45  045 -0.03834837  0.08659043
#> 46  046 -1.09571291 -2.02577998
#> 47  047  0.90189284  0.20950534
#> 48  048 -0.63572034 -2.05699728
#> 49  049  1.25331137  0.68799110
#> 50  050 -0.35368324  0.50141251
#> 51  051 -0.08088483  0.35392102
#> 52  052 -1.02019488 -0.84439879
#> 53  053 -0.33488895 -0.06158623
#> 54  054 -0.25742859 -0.01955451
#> 55  055 -1.62557142 -1.38926610
#> 56  056 -0.76427973 -0.41280528
#> 57  057 -1.04311700  2.00921531
#> 58  058 -0.26651118  0.32904265
#> 59  059  0.35512650  0.11252435
#> 60  060  0.12208987 -0.79917659
#> 61  061 -1.13598899  1.32732252
#> 62  062  0.59631295  1.11057287
#> 63  063  0.62714112  1.20044590
#> 64  064  0.71855980  0.82108776
#> 65  065 -0.57623218 -0.21643637
#> 66  066  1.65658819  0.60552122
#> 67  067 -0.09202413 -0.61846604
#> 68  068 -0.74765562  0.39017202
#> 69  069  0.10543574 -0.49757279
#> 70  070 -1.11471738 -0.27840541
#> 71  071  0.55440138  0.96837440
#> 72  072 -1.12296189  0.85298903
#> 73  073 -1.63345789 -0.51359747
#> 74  074  0.98070253  1.00087710
#> 75  075 -0.62303401 -0.75215649
#> 76  076 -0.64625163 -0.10702436
#> 77  077 -0.57940449 -0.33475019
#> 78  078  0.61469599 -0.78089957
#> 79  079  1.35159082  1.84913021
#> 80  080 -0.53806457 -0.85921172
#> 81  081 -1.06181874 -1.26669736
#> 82  082 -0.17974736  0.99516021
#> 83  083  0.85266337  1.29742904
#> 84  084 -0.86079216 -1.90365642
#> 85  085  0.91512661  0.42366332
#> 86  086  0.57909646  0.52842951
#> 87  087 -0.73093308 -0.24138912
#> 88  088 -0.11993975 -0.68640151
#> 89  089  0.23018132  0.36681619
#> 90  090  0.32226303  2.11494333
#> 91  091  1.01139769 -0.58403996
#> 92  092  1.23681573  0.13178363
#> 93  093  0.96224519  1.19667508
#> 94  094 -0.17908559  0.17210779
#> 95  095  0.65284702 -0.12272852
#> 96  096 -0.69101710  1.59737583
#> 97  097  0.28554277  0.51612978
#> 98  098 -0.57177771 -0.13672814
#> 99  099 -0.41125033 -0.72739363
#> 100 100  1.07458748 -0.32802852