Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5788573
#> Y_2 0.5788573 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.8692137  1.03824910
#> 2  0.4750551  0.16227917
#> 3  1.3827989  0.01977879
#> 4 -1.5113822 -1.48508311
#> 5 -0.1146402  0.15411981

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.979289896 -0.332324652
#> 2   002  0.617494538  0.003522683
#> 3   003  0.463038923  0.505687917
#> 4   004  0.549675622 -0.276054689
#> 5   005 -0.402483994  0.106999486
#> 6   006 -0.854864279 -0.712292026
#> 7   007  0.064964890 -0.248489854
#> 8   008 -0.044869453 -1.248430641
#> 9   009  0.911605047  0.463105697
#> 10  010  0.009351615  0.692354242
#> 11  011 -0.349710709 -0.538157678
#> 12  012 -0.159796125  0.831793378
#> 13  013 -0.437813843  0.073888617
#> 14  014  1.169556239  0.311274109
#> 15  015  0.267751040 -0.830909225
#> 16  016 -1.793712359 -1.490328681
#> 17  017 -0.795421191 -0.106059908
#> 18  018 -0.450832476 -1.817191558
#> 19  019 -0.600300455  0.657085898
#> 20  020 -0.252053216 -0.894558908
#> 21  021  0.452850380 -1.181712708
#> 22  022  0.352652359  1.285242574
#> 23  023 -1.754950272  0.058543493
#> 24  024  0.062078604 -0.591783963
#> 25  025 -0.568175892 -0.219818819
#> 26  026 -0.194179819 -0.891042249
#> 27  027 -0.974277512 -0.378101638
#> 28  028  0.092473099  1.147676652
#> 29  029 -0.388481826 -0.956801749
#> 30  030  0.538051300 -0.623368104
#> 31  031  0.450436678  0.056918035
#> 32  032  1.030603010  0.982150596
#> 33  033 -1.366691247  0.174663200
#> 34  034  0.140607678  0.239987729
#> 35  035  1.864287305  2.766624773
#> 36  036  0.463904166  0.527021325
#> 37  037 -0.612704311 -0.234137902
#> 38  038  1.140100038 -0.077829448
#> 39  039  1.092224345 -0.074084960
#> 40  040 -0.114326533  0.452756826
#> 41  041  1.333564845 -0.466599861
#> 42  042 -1.176122130 -0.865243943
#> 43  043 -1.707846492 -1.192680203
#> 44  044  0.529853087  1.772274133
#> 45  045  0.977947496  0.487526930
#> 46  046 -1.696646754 -0.415392066
#> 47  047 -1.114262744 -1.010305035
#> 48  048 -0.557853648  0.496927572
#> 49  049 -0.535864352  0.169212607
#> 50  050  0.605487267  0.602070491
#> 51  051  0.322997036 -1.028119237
#> 52  052 -0.810163895 -0.104529422
#> 53  053 -0.382722362  0.332508765
#> 54  054 -1.054997323 -0.426886800
#> 55  055  0.415137357  0.126120607
#> 56  056  2.060890118  1.557703193
#> 57  057 -1.880592012 -0.495025574
#> 58  058  0.446517430  1.365622196
#> 59  059  0.069088967 -0.133999397
#> 60  060  0.597081381 -0.228532061
#> 61  061  0.548282756 -0.389977615
#> 62  062 -0.766214320 -0.594304451
#> 63  063 -0.037831010  0.629140066
#> 64  064  1.420256143  0.558046913
#> 65  065  0.202667920 -0.600494626
#> 66  066  1.441543185  1.569095433
#> 67  067  0.801526927 -1.238515782
#> 68  068 -0.684548287  0.038332636
#> 69  069  0.099264265  0.306726564
#> 70  070 -0.620686164 -0.517320750
#> 71  071  1.154310490  0.711448611
#> 72  072  1.826583042  2.921990246
#> 73  073 -2.091984344 -1.528435816
#> 74  074 -1.087298083 -1.562347568
#> 75  075  1.585062124 -0.149439471
#> 76  076 -0.972284037 -1.108455513
#> 77  077 -0.760802958 -0.519800159
#> 78  078  1.332311044 -0.001899707
#> 79  079  0.677297290 -0.002897410
#> 80  080 -0.643179546 -0.058453167
#> 81  081  0.175838386  0.554632503
#> 82  082 -0.966779058 -1.388679464
#> 83  083  1.613322503 -0.551777462
#> 84  084  1.197322839  0.086523648
#> 85  085 -0.914744875 -1.830771680
#> 86  086  0.326425844 -1.347890019
#> 87  087  0.622610888 -1.422412521
#> 88  088 -0.536122518  0.014810591
#> 89  089 -1.612340425 -0.246826543
#> 90  090  0.440893385  0.652210695
#> 91  091  0.152555635  0.091400250
#> 92  092  0.810444778  0.132731655
#> 93  093  0.968368306 -0.337994783
#> 94  094  0.231939557 -0.456558235
#> 95  095 -0.733263167 -1.278670153
#> 96  096  1.649388049  1.447666971
#> 97  097  1.047226409 -0.033737517
#> 98  098 -1.038487893 -0.229126319
#> 99  099  1.081361028  0.750865163
#> 100 100  0.338612552  0.347104311

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001  0.58787334  0.1534925973
#> 2   002  2.22771787  3.3496087761
#> 3   003  2.44105713  0.7395538118
#> 4   004  2.38854219  2.7294156781
#> 5   005 -0.14353793  0.3647354904
#> 6   006  1.79068245  1.7438021351
#> 7   007  0.28813016 -0.3639110687
#> 8   008 -1.51010355 -1.3125299747
#> 9   009  0.23785426 -1.1183334797
#> 10  010  0.25409358  0.9708422615
#> 11  011 -0.60070833 -0.2540309563
#> 12  012 -0.97099636 -1.5616150647
#> 13  013  2.07556541  1.1153542485
#> 14  014 -1.64237024  0.0988086826
#> 15  015  1.74872713  1.5086764091
#> 16  016  1.44960572 -0.5505864799
#> 17  017  0.12469924 -0.1299958781
#> 18  018  0.01638799 -0.7399654392
#> 19  019  1.02119545  0.4618001251
#> 20  020  0.98523422 -0.1483289075
#> 21  021  1.24668655  1.2348760520
#> 22  022  1.07168603  0.7843468880
#> 23  023  0.37024677  1.1823798966
#> 24  024  0.02013667  0.6846005812
#> 25  025  0.63732652  1.5233683620
#> 26  026  0.94137642  0.0768907635
#> 27  027 -1.30863620 -2.9930174491
#> 28  028 -0.79742559 -1.4995173385
#> 29  029 -2.44700341 -1.7223906967
#> 30  030 -1.07935452 -0.5155320586
#> 31  031 -0.60649122 -1.4011664504
#> 32  032 -0.50343117  0.0554588087
#> 33  033  0.26028372 -0.8999261077
#> 34  034 -0.77205000  0.9575861268
#> 35  035  0.87847690 -1.2384923028
#> 36  036  0.82130507 -0.1094950629
#> 37  037  0.68923186 -0.5710316310
#> 38  038  0.67752056 -1.1270381921
#> 39  039  0.85652773  0.3827844920
#> 40  040 -0.40298131 -0.7060650680
#> 41  041 -1.75502495  1.2505168691
#> 42  042  0.30417688  0.4651500159
#> 43  043 -2.02033506 -1.5397705687
#> 44  044 -0.87698591  0.9152031297
#> 45  045  0.52195483  0.1253840410
#> 46  046  0.86804168  1.2751169469
#> 47  047 -1.77196747 -1.2123124744
#> 48  048  0.98572945 -0.1462201599
#> 49  049 -1.51896385 -0.5729984548
#> 50  050 -0.10328488 -0.2865219060
#> 51  051 -1.29534216 -0.2591168469
#> 52  052 -1.18348979  0.1664667345
#> 53  053 -1.35803369  0.0229712905
#> 54  054  1.06411179  1.4067704652
#> 55  055 -1.05361040 -1.3628154967
#> 56  056 -0.14734134 -1.9366046223
#> 57  057  0.48360697 -0.4085658587
#> 58  058 -0.31704950  0.1743153105
#> 59  059  0.69003290 -0.2726654786
#> 60  060  0.75003573  1.1399145345
#> 61  061 -1.17281258 -1.5006876583
#> 62  062  0.76014940 -0.4756308608
#> 63  063  0.58838647  0.0269350460
#> 64  064 -0.98435952 -0.5841932975
#> 65  065 -1.12258389 -2.2480083457
#> 66  066 -0.43134479 -0.0509004930
#> 67  067  0.36263029 -0.0219973776
#> 68  068  0.11572747  0.6158928588
#> 69  069  0.34525118  0.9318186727
#> 70  070  1.69760365  1.7727830144
#> 71  071 -1.45222315 -1.1494219303
#> 72  072 -1.53937728 -0.3636738488
#> 73  073  0.48984766 -0.9260196182
#> 74  074 -0.85505748 -2.6320362735
#> 75  075  0.40895259  0.7963635110
#> 76  076  1.82281362  0.0001435941
#> 77  077  0.79531075  1.3419330948
#> 78  078  0.15563409  0.0077390891
#> 79  079  0.58527665  0.9693376805
#> 80  080  1.23227091  1.7759995661
#> 81  081 -2.04460619 -0.7133888083
#> 82  082 -0.93278611 -1.7721844043
#> 83  083  0.37362868  0.9266054142
#> 84  084  0.08957723  0.9461852923
#> 85  085  0.43555571  2.1559223116
#> 86  086 -1.15802612 -0.3850201444
#> 87  087 -0.52686529 -1.0142058589
#> 88  088  0.34496409  0.0212329027
#> 89  089 -0.18161811  0.3426344572
#> 90  090  0.69012687 -1.0518397215
#> 91  091 -1.28347109 -0.9733180499
#> 92  092 -0.35600368 -0.3704296394
#> 93  093  1.93696121  1.5789631416
#> 94  094 -0.21829577 -1.6325574708
#> 95  095 -0.54901178 -1.6061494029
#> 96  096  1.13622039  0.5694310557
#> 97  097 -0.70658625  0.3706395593
#> 98  098 -2.13283199 -0.8265649689
#> 99  099 -0.87848873 -0.3058166052
#> 100 100  2.21741064  1.3084939076