Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4538386
#> Y_2 0.4538386 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5800178  1.4860630
#> 2 -0.2625864  0.1493868
#> 3  2.0762674 -1.0288211
#> 4 -2.0501768 -0.5919241
#> 5 -1.2914684 -1.1953211

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1          Y_2
#> 1   001  0.9639964850  0.843414468
#> 2   002  0.9910458231 -0.150033845
#> 3   003 -0.3622205418  1.710837451
#> 4   004 -0.1548589848 -1.236548971
#> 5   005 -0.4778545707 -0.999076798
#> 6   006 -0.9710280889 -1.346441659
#> 7   007 -1.3643508733 -1.013602670
#> 8   008  0.4341192088  1.220458931
#> 9   009  0.0777821156 -0.417009044
#> 10  010 -0.3934154756  0.187288487
#> 11  011 -0.8309346500 -0.526061089
#> 12  012  0.5317574526  0.044606380
#> 13  013  0.4391415457  0.067391843
#> 14  014  1.3561748665  0.661879959
#> 15  015 -0.2101405351 -1.421129716
#> 16  016  1.1155326398  0.461357069
#> 17  017  0.6304181688  0.694320369
#> 18  018 -1.5558161211 -0.700319531
#> 19  019  0.5682895600 -1.283325246
#> 20  020 -0.4957800913 -0.446252505
#> 21  021  0.2454165117 -0.667301707
#> 22  022 -0.4667838915 -0.907705864
#> 23  023 -0.2806961004 -0.116829316
#> 24  024  0.1519784879  0.862107122
#> 25  025  1.3078800238  2.382445566
#> 26  026  1.9650167769  2.298445593
#> 27  027  0.2019281603 -1.174663660
#> 28  028  0.1569142542  0.584692667
#> 29  029 -0.7439940627  1.147704613
#> 30  030  1.6262777278  0.707025639
#> 31  031 -1.0867763265 -1.569953462
#> 32  032 -1.0038523997 -1.686138841
#> 33  033 -0.4399940368 -0.928603215
#> 34  034 -0.1036846044 -0.570879502
#> 35  035  0.4807149827 -0.926123875
#> 36  036 -0.7248309708  0.989284492
#> 37  037 -0.1486260159 -0.962635554
#> 38  038 -0.6392144484 -0.195925724
#> 39  039 -0.3721913751  0.195627582
#> 40  040  0.0797250228 -0.234855676
#> 41  041  0.6979108603  1.680697151
#> 42  042 -0.8172947637 -0.305933701
#> 43  043  0.3304572814  1.710063029
#> 44  044 -0.2430479951 -1.328724314
#> 45  045 -1.0686387639 -0.372564663
#> 46  046  0.3972056756 -0.157303901
#> 47  047 -1.7249313822 -1.410000268
#> 48  048 -0.3852406117 -0.007873116
#> 49  049 -0.3326270669  2.116874333
#> 50  050  0.3018880567 -0.904906541
#> 51  051 -0.3278306493  0.171425089
#> 52  052  0.6641407500  0.685866919
#> 53  053 -0.2862136946 -0.349047515
#> 54  054 -0.1622055027 -0.687991883
#> 55  055 -1.1457455061  0.328326593
#> 56  056  0.8057340985 -1.012132880
#> 57  057  0.7387635011  1.081539154
#> 58  058 -0.2456973433 -1.381791608
#> 59  059 -1.2802071618 -1.697847121
#> 60  060 -1.2163109911 -1.773109867
#> 61  061 -0.3057175834 -1.169245228
#> 62  062 -0.2324536885  0.798691878
#> 63  063  2.7039306132  1.777641554
#> 64  064  1.3021006700  0.743079680
#> 65  065  0.7021694400  1.040899503
#> 66  066  2.4826819774  2.291562615
#> 67  067  0.0134918103  0.211977273
#> 68  068 -0.1496632459  0.296986823
#> 69  069 -0.5692460032  1.415200373
#> 70  070 -0.1204587653 -0.002766311
#> 71  071  1.3717393570  1.863579609
#> 72  072 -0.3750528017 -0.289510217
#> 73  073 -0.8507042088 -1.306988592
#> 74  074 -1.1288905198 -1.383588836
#> 75  075 -1.0822742395 -0.980177193
#> 76  076  0.0913533890  0.024813972
#> 77  077 -0.0003152268 -0.260217463
#> 78  078 -1.2067407269 -0.565224391
#> 79  079  0.0149626379 -0.661301647
#> 80  080  0.4983515789  0.225525762
#> 81  081  0.2955322347 -1.365879428
#> 82  082  1.7280617005  0.886682172
#> 83  083  0.3668501492 -0.413763737
#> 84  084  0.5642503254  0.889281093
#> 85  085  1.7088859541 -0.674317510
#> 86  086  0.9126100942  0.931176969
#> 87  087  1.3780894874  0.848002821
#> 88  088  0.2481683799  0.247099385
#> 89  089 -0.6911012420  0.415541428
#> 90  090  0.0503484239 -0.140042439
#> 91  091 -1.5404039956 -1.669007559
#> 92  092  0.3883122453  0.145460430
#> 93  093 -1.0761056885 -0.103900564
#> 94  094 -0.2913570027 -0.864689641
#> 95  095  0.5447942521 -0.147004278
#> 96  096 -0.1768111389 -0.614842744
#> 97  097 -0.5151410940 -0.774103203
#> 98  098 -0.6901589806  1.145435724
#> 99  099  0.2393529841 -0.309214599
#> 100 100 -1.4516620387 -1.510179047

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.470595806  0.44538949
#> 2   002  1.290177742 -0.83781241
#> 3   003 -0.473236336  0.39594966
#> 4   004 -0.195447957  1.65534387
#> 5   005 -0.880858417 -0.89239558
#> 6   006  1.011102467  0.80990468
#> 7   007  1.661161528 -0.93967763
#> 8   008  0.660919215 -0.92677614
#> 9   009 -0.681099304 -0.80211276
#> 10  010 -0.375243138 -0.85975877
#> 11  011  0.565228688  0.53365285
#> 12  012  1.143164154  0.40543678
#> 13  013 -0.527542310 -0.43921874
#> 14  014 -1.115439912  0.13840516
#> 15  015  0.912806565  0.02528008
#> 16  016  0.640145175 -0.49575711
#> 17  017  0.565102428 -0.31003746
#> 18  018  0.594370300  1.31382437
#> 19  019  1.421618770  1.24190460
#> 20  020 -1.159500876 -0.38650619
#> 21  021  1.297380641  2.31209445
#> 22  022  0.204727318  0.75628199
#> 23  023  0.211947214 -0.10563641
#> 24  024 -1.882215778 -1.62625548
#> 25  025  0.896757153  0.06663088
#> 26  026 -0.336286324 -0.12163323
#> 27  027  0.215486345  0.39770632
#> 28  028 -1.357888106  0.68170211
#> 29  029 -2.481780898 -1.86345502
#> 30  030  0.540823218  0.53944004
#> 31  031  0.408532029 -0.98473321
#> 32  032 -0.311857597 -0.24920459
#> 33  033 -0.503029328  0.49961462
#> 34  034 -0.322728090  1.03750480
#> 35  035  0.286591517  0.31958523
#> 36  036  0.697183488 -0.39854148
#> 37  037  0.493948148 -0.64288024
#> 38  038 -2.301159122 -2.74436780
#> 39  039  0.868151557  1.53000550
#> 40  040  1.408964098  2.52788721
#> 41  041 -0.117977676 -1.49663695
#> 42  042  1.437558122  1.52352206
#> 43  043 -0.645622090 -0.04539906
#> 44  044  1.254233336  1.24500058
#> 45  045 -2.064669712 -1.58724029
#> 46  046  0.453009690  0.68553420
#> 47  047 -2.000037816 -1.15157685
#> 48  048 -0.864643548  0.27227826
#> 49  049  0.252733870  0.26260593
#> 50  050 -0.782123638 -0.06743613
#> 51  051  0.605757342  0.64373551
#> 52  052 -1.765057801 -0.80560517
#> 53  053  1.085296431  0.24416919
#> 54  054 -1.174554363 -1.60849153
#> 55  055  0.628579109 -0.08880857
#> 56  056 -0.356023414 -1.59821600
#> 57  057  0.693143970 -0.65163441
#> 58  058 -0.225027737 -0.82313050
#> 59  059 -1.030398017  0.27223621
#> 60  060  0.410219664  0.07762842
#> 61  061  0.567477796  0.36943452
#> 62  062 -1.898125754 -0.19474055
#> 63  063 -0.967087044 -0.75173395
#> 64  064  0.064403213 -1.07795350
#> 65  065 -1.008403322  1.16592952
#> 66  066  0.859055435 -0.90143336
#> 67  067 -0.151776197 -0.41954068
#> 68  068 -0.286560150  0.93345843
#> 69  069  0.002576576 -0.35913511
#> 70  070 -0.852886777 -0.75733893
#> 71  071 -0.078899947 -1.14380476
#> 72  072  0.697804174  1.41527163
#> 73  073  1.494298543  0.01272732
#> 74  074  0.702895758 -0.02186542
#> 75  075  0.658300314  0.76513535
#> 76  076 -1.589308533  0.19061413
#> 77  077 -0.445516935 -0.12607905
#> 78  078  0.325965981  0.72432983
#> 79  079  1.182357632 -0.26084260
#> 80  080 -0.877934969 -1.67515918
#> 81  081 -0.032232018 -1.27964120
#> 82  082  1.013666750 -0.47831068
#> 83  083 -0.825030445 -0.19609199
#> 84  084  0.150796803 -0.24038097
#> 85  085 -0.350894958 -0.68579904
#> 86  086 -0.016286664  0.27430034
#> 87  087 -0.880383861 -0.73713766
#> 88  088  0.175723170  1.42496197
#> 89  089 -0.180569071 -0.27917512
#> 90  090  1.260652457  1.14869358
#> 91  091 -0.831000543 -1.99802735
#> 92  092 -1.571475927 -1.65045004
#> 93  093  0.813045760  0.83548740
#> 94  094  1.184384200  0.63033745
#> 95  095  0.551239239 -1.36831632
#> 96  096  1.599760320  1.40733699
#> 97  097  1.547027781 -0.67965895
#> 98  098  0.145569330 -1.55991648
#> 99  099  1.045179606 -0.37017284
#> 100 100 -0.278540603 -0.33453659