Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5090629
#> Y_2 0.5090629 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1          Y_2
#> 1 -0.2318510  1.058686491
#> 2 -1.5110553 -1.556443421
#> 3  1.9383432  2.412845812
#> 4  0.7066879  0.007487284
#> 5 -0.2553440 -0.281100806

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.269514547 -0.072343448
#> 2   002 -0.561245125 -0.359287515
#> 3   003  0.859400191 -0.372586687
#> 4   004  0.273865941  0.580131402
#> 5   005  0.943488911  0.518412889
#> 6   006  0.518669839  0.412437557
#> 7   007  0.978585024 -0.606559794
#> 8   008  0.947947844  0.683832270
#> 9   009 -1.341828923 -0.893417883
#> 10  010  0.842929372 -0.526986787
#> 11  011 -0.408077463  0.370487053
#> 12  012  1.791696339  0.389891650
#> 13  013  2.232232496  0.937983905
#> 14  014 -0.814977899 -0.226832805
#> 15  015 -1.369190237 -1.754819809
#> 16  016 -0.038305534  0.460815145
#> 17  017  0.562252310  2.421554452
#> 18  018  0.683659791 -0.082868901
#> 19  019  0.367208723  0.430943443
#> 20  020  1.512638559  1.152309631
#> 21  021  0.066076953 -0.393557145
#> 22  022  0.072001480 -0.022671601
#> 23  023  1.182235010  1.696047856
#> 24  024 -1.619013962 -0.753558591
#> 25  025 -0.439188645 -1.093371948
#> 26  026  0.402948606  0.979011587
#> 27  027  0.790944328  1.163299385
#> 28  028 -0.284506115 -0.556072226
#> 29  029  0.244834504  0.013554760
#> 30  030 -1.185440255  0.233733460
#> 31  031 -0.687241514 -0.370673270
#> 32  032  0.756335014 -0.010857178
#> 33  033 -0.563673133 -0.019493803
#> 34  034  1.183551018  0.413078751
#> 35  035 -0.608718756  0.847887282
#> 36  036  0.179250895  1.064307164
#> 37  037  1.896958503  1.399057937
#> 38  038  1.065976642  0.899184264
#> 39  039 -1.062674294  0.321353120
#> 40  040  0.763579969 -0.069330881
#> 41  041  1.969257216  1.496251574
#> 42  042 -0.869676036 -0.664490364
#> 43  043  0.269748668  0.411795156
#> 44  044  1.640161115  1.449647663
#> 45  045 -0.697083967 -0.437754264
#> 46  046 -1.178048989 -0.637543423
#> 47  047  2.069096330  0.808127042
#> 48  048  0.665064288  0.097706822
#> 49  049  1.818297199  0.392495372
#> 50  050  0.347927131  0.168407882
#> 51  051 -0.309564937 -0.528088440
#> 52  052  1.784441281  1.532137328
#> 53  053  2.873211526  1.414368508
#> 54  054  0.587416927 -0.222533602
#> 55  055 -0.133073151  1.496498087
#> 56  056  1.691682279  0.088667081
#> 57  057  1.682969101 -0.841854370
#> 58  058 -1.302712707 -0.579648208
#> 59  059  0.746369827 -0.977904225
#> 60  060  0.322868914  0.258194415
#> 61  061 -1.404472419 -1.746957239
#> 62  062  1.701427421  0.766250351
#> 63  063 -0.372765930 -1.696514689
#> 64  064 -0.596242121  0.180805399
#> 65  065  0.687291061  1.036588052
#> 66  066  0.556190218  0.328260819
#> 67  067 -0.265911544  0.882039832
#> 68  068 -0.234125647  0.503032246
#> 69  069  0.311712870 -0.757166035
#> 70  070  0.966635834  0.845849954
#> 71  071  2.283949998  0.635391240
#> 72  072 -0.381024624 -0.701276417
#> 73  073 -0.432469422  0.076031409
#> 74  074 -1.323694026 -1.005017077
#> 75  075 -0.208399904  0.150593533
#> 76  076  0.538599270 -0.638672127
#> 77  077 -0.125896379 -0.228500906
#> 78  078  0.008459032  0.228958585
#> 79  079  1.217501457 -1.329687966
#> 80  080 -1.251483335 -0.845054461
#> 81  081 -0.337058649 -0.198210529
#> 82  082  0.261031456 -0.817608917
#> 83  083  0.389962210  0.862701521
#> 84  084 -1.388443694  0.394667817
#> 85  085 -0.549757171 -1.869475380
#> 86  086 -0.458846673 -0.518223446
#> 87  087 -1.959860690 -0.957337311
#> 88  088 -2.246843526 -1.546091857
#> 89  089  0.195947322  0.006948101
#> 90  090 -0.514058554 -1.078017689
#> 91  091  0.505367719 -0.598477949
#> 92  092  0.369607790  1.740508615
#> 93  093 -0.830855941  0.399161285
#> 94  094 -0.891775329 -0.962727155
#> 95  095  0.261229414 -1.027226080
#> 96  096 -0.002108251  0.365599239
#> 97  097 -0.010185072 -2.142748967
#> 98  098  0.285690223 -0.031751311
#> 99  099  1.446337088  0.606505185
#> 100 100  3.112046508  0.371306327

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.24204970 -0.69005599
#> 2   002 -0.42006680 -0.59994328
#> 3   003  0.57454410 -0.65183057
#> 4   004 -0.27439524  0.12696355
#> 5   005 -0.17566561 -0.04310853
#> 6   006 -0.30003358 -0.80305787
#> 7   007 -0.19556651 -1.50214352
#> 8   008  0.72386818 -0.54907895
#> 9   009  0.23123409  0.63502574
#> 10  010  1.12889452  0.66175249
#> 11  011  0.69984002  0.40700190
#> 12  012 -0.35528487  1.39119940
#> 13  013 -1.81508488  0.36732925
#> 14  014  0.19864055 -0.26853663
#> 15  015 -0.78768555 -0.53332852
#> 16  016 -0.91230765 -0.81311402
#> 17  017  0.04524381 -0.55521294
#> 18  018  0.32024227 -0.62508248
#> 19  019  1.47042481  1.31556979
#> 20  020  1.64353939 -0.82422323
#> 21  021  1.49149057  1.39869527
#> 22  022  0.83471241  0.08545288
#> 23  023 -0.75881959 -0.13209684
#> 24  024 -0.55190569  0.38534091
#> 25  025  0.88654369  0.63004164
#> 26  026 -0.28031737  0.71587704
#> 27  027  2.30233592  3.00593175
#> 28  028 -0.02760762 -0.41752278
#> 29  029 -0.25980184 -0.64160739
#> 30  030  0.58593288  0.62856241
#> 31  031  0.06710968  0.37805647
#> 32  032 -0.89678086 -0.40289104
#> 33  033 -0.33860635  0.41673715
#> 34  034  0.56830920  1.91504369
#> 35  035  0.55288583  1.74431887
#> 36  036  0.87748131  1.16851130
#> 37  037 -0.10689133  0.03615757
#> 38  038 -1.08776772 -1.50889627
#> 39  039  0.59569477  1.28821718
#> 40  040 -1.16954839 -1.26094497
#> 41  041  0.17623986 -1.53862887
#> 42  042  1.08165637  0.72827578
#> 43  043  0.68542640 -0.36487244
#> 44  044  0.16669336  0.33172354
#> 45  045 -1.08112003 -0.14971503
#> 46  046  0.10591303 -0.50829446
#> 47  047 -0.71262670  0.22562461
#> 48  048 -0.42817740 -0.41127286
#> 49  049 -0.03528192 -0.15216070
#> 50  050  0.84003540  0.76244938
#> 51  051 -0.85681724 -0.72487637
#> 52  052 -0.86851017 -1.86179180
#> 53  053  1.17401166  1.34179559
#> 54  054 -0.04990040  0.11770143
#> 55  055 -0.34660554 -0.66182565
#> 56  056  1.97625370  0.66548894
#> 57  057 -0.52406250 -1.31547669
#> 58  058 -0.26996964 -1.08155539
#> 59  059  1.23761950  0.16082523
#> 60  060  2.27352181  0.45404618
#> 61  061  0.92293332  1.48113981
#> 62  062 -0.17030193 -0.40191597
#> 63  063 -0.64386846 -0.37974121
#> 64  064  1.39607287  0.55801064
#> 65  065 -1.05320927  0.39157160
#> 66  066 -1.41742995 -0.23384115
#> 67  067 -0.95394857 -1.79247704
#> 68  068  0.07491307 -0.12459901
#> 69  069 -0.56459168 -0.49323336
#> 70  070 -0.21188092  0.38438155
#> 71  071 -0.52612930  0.48956394
#> 72  072  0.25913679 -0.46026524
#> 73  073 -0.21885279  0.80395873
#> 74  074  0.64989955  0.56374666
#> 75  075 -0.41324803  0.23186947
#> 76  076 -0.90116946 -1.27180414
#> 77  077  0.39822010  0.48261963
#> 78  078  0.36332159  0.43475220
#> 79  079  1.04239996  0.53232201
#> 80  080 -1.97853623 -1.62097322
#> 81  081 -0.60530986  0.55624765
#> 82  082  0.72063279  0.76200973
#> 83  083 -1.43312195 -0.75954149
#> 84  084 -1.03875218 -0.55652383
#> 85  085 -0.45996864 -1.03244421
#> 86  086  0.94819430 -0.22502700
#> 87  087  0.25862903 -0.18838484
#> 88  088  0.59425928  1.23300117
#> 89  089 -0.05227777 -0.36723877
#> 90  090 -0.10128220 -0.99122352
#> 91  091 -1.73772380 -0.90171631
#> 92  092  1.49910876  0.96129517
#> 93  093 -0.22994242 -0.28477690
#> 94  094  1.38151483 -0.01137876
#> 95  095 -0.94964084  1.63503154
#> 96  096 -1.40502452 -0.37218422
#> 97  097 -0.03233207 -0.68010456
#> 98  098  0.56882728 -0.57510739
#> 99  099 -1.05865323 -1.40664358
#> 100 100  0.77141993 -0.49885066