Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5055433
#> Y_2 0.5055433 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.5785533  0.30930244
#> 2  0.6061655  0.06497496
#> 3  0.5236274 -0.22631563
#> 4 -0.6731935 -0.23517511
#> 5 -0.8083438 -0.82844304

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.69107150  1.67143574
#> 2   002  1.64262873  1.13810969
#> 3   003  0.41523859  0.83103323
#> 4   004 -0.73631703 -0.96188190
#> 5   005  1.01420075 -0.04347429
#> 6   006  0.03201250 -0.04424299
#> 7   007 -0.40841093  0.19291364
#> 8   008  0.95697751 -1.50150222
#> 9   009 -0.25123261  0.34075347
#> 10  010 -0.18776770 -0.61531329
#> 11  011  0.77538530 -0.15399097
#> 12  012  0.12474084  1.05493257
#> 13  013  0.28234316  0.67493945
#> 14  014 -0.35273525 -1.16190564
#> 15  015  0.67325619  1.64193867
#> 16  016 -0.50136583  1.25696393
#> 17  017  1.12203751  1.44319537
#> 18  018  1.27690084  0.35146065
#> 19  019 -1.45009640 -0.47781622
#> 20  020 -1.32265529 -1.30009418
#> 21  021 -1.10575492  0.72631358
#> 22  022 -0.01001446 -0.56758580
#> 23  023  1.73559793  0.91679394
#> 24  024 -2.99959566 -1.95777694
#> 25  025  0.75318555 -0.90712435
#> 26  026  1.37159408  0.46355481
#> 27  027 -0.08628178 -1.35921772
#> 28  028  0.06920038 -1.26564255
#> 29  029  0.41852802 -2.08273160
#> 30  030  2.29126652  0.56781391
#> 31  031 -1.30838168 -1.52578550
#> 32  032 -0.65616922 -0.25210429
#> 33  033 -1.45126226 -0.79294815
#> 34  034 -1.48472614  0.13448399
#> 35  035 -0.57590320  0.70480382
#> 36  036 -0.89505788  0.04129825
#> 37  037  0.36897173  1.60512771
#> 38  038  0.12987742  1.58436174
#> 39  039 -1.88060813 -1.39489881
#> 40  040  0.14925447 -0.06278967
#> 41  041  1.01941397  2.34456107
#> 42  042  1.64676320 -0.48260222
#> 43  043  1.16492577  0.46970709
#> 44  044  0.69177206 -0.34162840
#> 45  045  0.78843167  0.16861723
#> 46  046 -0.91854110 -0.36389184
#> 47  047 -1.41616146 -0.38863309
#> 48  048 -0.24775814 -1.53736202
#> 49  049  0.51157674  0.29254233
#> 50  050  1.61109403  0.08221296
#> 51  051  0.82288246  0.57411154
#> 52  052 -0.78329198 -0.76673395
#> 53  053  2.35827185  0.94816511
#> 54  054 -0.37696909 -0.81707064
#> 55  055  0.99629094 -0.65333300
#> 56  056  1.14383946 -0.11216517
#> 57  057  1.08335799  0.44347566
#> 58  058 -0.75456837  0.07509377
#> 59  059 -1.52576917 -1.44043659
#> 60  060  0.43734116  0.74716839
#> 61  061 -0.14818671  0.11934736
#> 62  062  0.57945116 -1.07743424
#> 63  063 -1.44897547  0.99591717
#> 64  064  1.30028130 -0.39507101
#> 65  065  0.44246520  1.44945507
#> 66  066  0.74051389  1.86124772
#> 67  067 -0.66918061 -1.17772504
#> 68  068 -0.80926380 -0.82238167
#> 69  069 -1.71304308 -1.52527361
#> 70  070  0.95540300 -0.18672985
#> 71  071  0.65128512  0.23768235
#> 72  072 -0.66918769  0.42128507
#> 73  073  0.62506779  0.32681488
#> 74  074  0.97342128  1.02516587
#> 75  075 -0.77926419 -0.67258924
#> 76  076  0.71803933  0.81506107
#> 77  077  1.02097490  0.46158313
#> 78  078  0.03747044 -0.38583452
#> 79  079 -0.45144045 -0.22381795
#> 80  080  0.51374600  0.19743127
#> 81  081 -1.86167031  0.11038375
#> 82  082  0.95610487  1.73368841
#> 83  083 -1.72786678 -0.39802146
#> 84  084 -0.43602127  0.01901590
#> 85  085  0.73081641  0.71650094
#> 86  086  1.41578167  0.66975290
#> 87  087  0.83546906 -0.74469084
#> 88  088 -1.86114688 -0.96289323
#> 89  089 -0.43462888  0.60131521
#> 90  090 -0.45313604  0.28817315
#> 91  091  0.16867422 -1.98197911
#> 92  092 -0.94732511 -1.11256268
#> 93  093 -0.93019278 -1.92966300
#> 94  094  0.03258731 -0.88316955
#> 95  095  0.93727717  0.37176441
#> 96  096  1.72542719  2.29758100
#> 97  097  1.36092149  0.05152730
#> 98  098  0.82500014  0.40425699
#> 99  099  1.30664588 -0.90651147
#> 100 100 -0.80136880  1.50362216

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.429476524  0.106062400
#> 2   002  0.315861610 -0.036390895
#> 3   003  0.705215456 -0.671813815
#> 4   004  1.252321898  0.070673572
#> 5   005 -0.760019632 -0.297393018
#> 6   006  1.040582088  2.063595349
#> 7   007  1.072462360  0.824538816
#> 8   008  2.741262794  0.497242135
#> 9   009  0.257270344  0.299525776
#> 10  010 -1.810832989 -0.041187922
#> 11  011 -0.059866509 -1.189879606
#> 12  012 -0.913804040 -0.277109062
#> 13  013  1.212311179  0.439413699
#> 14  014 -0.829893039 -1.428651657
#> 15  015  1.985762380  2.058385782
#> 16  016 -0.161157362  1.633513636
#> 17  017  0.999687604  0.244021003
#> 18  018 -1.607941756 -1.536825219
#> 19  019  1.499479891  0.647039593
#> 20  020  0.362614167  0.265124876
#> 21  021 -0.511143951 -0.346241923
#> 22  022  1.135134668  0.230800844
#> 23  023  0.021046846  0.437878837
#> 24  024  0.667382232  0.732003610
#> 25  025 -0.046694411 -2.034704839
#> 26  026  0.051051863  0.975908697
#> 27  027 -1.598828367 -1.497905657
#> 28  028 -1.986530734 -1.630982793
#> 29  029 -0.922576102 -0.886833468
#> 30  030  0.558452071  2.510025281
#> 31  031 -0.072063797  2.121862935
#> 32  032  0.327111350  1.330361724
#> 33  033 -1.385306409 -0.421514072
#> 34  034 -0.647701477 -1.821312317
#> 35  035 -0.554581851  0.653748722
#> 36  036 -1.209415336  0.211311700
#> 37  037 -1.166083314 -1.512643869
#> 38  038  0.612084865  0.818720523
#> 39  039 -1.004693514 -1.266943017
#> 40  040 -0.331073830 -0.513043282
#> 41  041 -0.349295978  2.164482703
#> 42  042 -1.902172719 -2.671028531
#> 43  043  0.174274481  0.815419466
#> 44  044  1.158707554  1.081388476
#> 45  045  0.149760566  0.853171194
#> 46  046  0.206141612 -0.040133669
#> 47  047  0.792569680  0.260073389
#> 48  048 -1.169110214 -1.522450793
#> 49  049  0.961916810  0.225149955
#> 50  050 -0.477088904  0.106083719
#> 51  051  1.003475492  1.811606192
#> 52  052  0.008295952  0.353626349
#> 53  053 -2.895678131 -2.273826661
#> 54  054 -1.048296526 -0.507203671
#> 55  055  0.653221381  0.240060272
#> 56  056  0.653636882 -0.302320669
#> 57  057 -0.132396977  0.479174145
#> 58  058 -1.347289686 -0.201835094
#> 59  059 -1.412027818  0.609469114
#> 60  060  0.361355503  1.919072643
#> 61  061  0.352297616  0.843205290
#> 62  062 -0.318748373 -0.221914780
#> 63  063 -0.685689567 -0.803377298
#> 64  064 -2.756156549  0.331331533
#> 65  065  0.553986097  0.424240911
#> 66  066  0.204474517  0.050168439
#> 67  067  0.689112622  0.717779358
#> 68  068 -0.692513162 -1.449197993
#> 69  069 -0.674889626 -0.545417646
#> 70  070  0.363868397 -0.770451366
#> 71  071 -1.882113644 -1.710935035
#> 72  072  0.107444831  0.479180972
#> 73  073  0.359993879  1.062148581
#> 74  074 -1.711735040  0.936174971
#> 75  075 -2.240754550 -0.922570546
#> 76  076  2.192543671  1.859380293
#> 77  077  0.064161476 -0.101417251
#> 78  078 -0.889429823  0.270970569
#> 79  079 -0.049774487 -1.119882166
#> 80  080 -0.019974983  1.598515540
#> 81  081 -0.680947216  0.958133273
#> 82  082  1.594155700 -0.187055489
#> 83  083 -0.211787030  0.004200879
#> 84  084 -0.673554976 -0.128605382
#> 85  085  0.111616284  0.661401720
#> 86  086 -0.446448966 -0.447948718
#> 87  087  1.071782491  0.703785438
#> 88  088  0.989566669  0.319731243
#> 89  089  1.303888231  0.420626507
#> 90  090  0.023531701 -0.265573938
#> 91  091 -0.500340959 -1.262574551
#> 92  092 -0.266789824  0.273797935
#> 93  093 -0.553950950  0.125387735
#> 94  094  1.678443308  0.775962514
#> 95  095  0.802838044 -1.113777719
#> 96  096  0.179186507 -0.208069684
#> 97  097 -0.217328208  0.584305006
#> 98  098 -0.411102977  0.027297318
#> 99  099 -2.550087768 -2.696008005
#> 100 100 -0.359679490 -0.622553690