Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4467477
#> Y_2 0.4467477 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2591715  0.8815865
#> 2 -0.9987356 -0.3058587
#> 3 -0.6357638 -1.3026463
#> 4  1.5069704 -0.6332418
#> 5 -0.4052699 -1.2754827

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.017035668  0.594442068
#> 2   002 -1.054374731 -0.407915301
#> 3   003  0.203425596  0.067688179
#> 4   004  0.186260184 -0.699678346
#> 5   005 -0.474816073 -0.965479166
#> 6   006 -1.066872820 -0.292739753
#> 7   007 -1.135885775  0.115857230
#> 8   008  0.833793002 -0.662367687
#> 9   009 -2.509572227 -0.238935428
#> 10  010 -1.187738579  0.397406600
#> 11  011 -0.784227850  0.832062146
#> 12  012  3.168131704  1.010840379
#> 13  013  1.231874122  0.366495416
#> 14  014  0.563968261  2.815518858
#> 15  015 -0.325488680 -0.546481996
#> 16  016  0.669996358 -0.303445976
#> 17  017 -1.686128263 -1.148746124
#> 18  018  0.468079175  1.539441615
#> 19  019 -0.065677179 -0.073929432
#> 20  020 -0.224343714 -0.914218212
#> 21  021  0.878280357 -0.720249606
#> 22  022 -0.067846189 -1.460194433
#> 23  023 -1.529951531 -1.006697180
#> 24  024  1.596270128 -0.061011505
#> 25  025  1.383717014  0.559797073
#> 26  026  0.102615873 -0.707288002
#> 27  027  0.617147709  1.439246308
#> 28  028 -1.817594065 -1.874602071
#> 29  029 -1.653811154 -1.219091313
#> 30  030  0.073324014 -0.479590152
#> 31  031 -1.809726997  0.244752821
#> 32  032  0.828623795 -0.976412910
#> 33  033  0.172451626 -1.672814283
#> 34  034  0.250660955  0.877726250
#> 35  035 -0.485728401  0.818446977
#> 36  036  1.070615943  1.805227365
#> 37  037 -0.953048048  0.286609448
#> 38  038  0.433446209  0.665486896
#> 39  039  0.476597914  1.314598430
#> 40  040  1.225568236  0.928382468
#> 41  041  1.967566728  0.597180236
#> 42  042 -1.353563255 -0.088400773
#> 43  043 -0.356171555  0.342434921
#> 44  044 -0.082720562  0.676579589
#> 45  045  0.829354171  0.328918778
#> 46  046 -0.549715591  1.237770587
#> 47  047  0.445296272  0.390009233
#> 48  048 -0.936157994  0.080229413
#> 49  049 -0.390653379 -0.644179682
#> 50  050  1.024284895  0.674891378
#> 51  051 -0.019660024 -0.866723560
#> 52  052  1.480193932  0.738562980
#> 53  053 -0.359531421  1.218670625
#> 54  054  0.528718628  0.376249167
#> 55  055 -0.216772312 -0.865856197
#> 56  056  0.953277446  1.094205100
#> 57  057 -0.430888289 -0.219785680
#> 58  058 -0.208988864 -1.600093171
#> 59  059 -0.737755429 -1.605805613
#> 60  060  0.626677077 -0.175725987
#> 61  061 -1.197578181  0.900964944
#> 62  062 -0.637976253 -0.042124099
#> 63  063 -1.336863899 -0.827139365
#> 64  064 -0.894468620 -0.236717332
#> 65  065 -1.104700588  0.255544068
#> 66  066  0.222632235  0.225566282
#> 67  067 -1.428239503 -1.636152522
#> 68  068 -2.377460646 -0.994174526
#> 69  069  0.846830654  0.982494794
#> 70  070 -2.987133681 -1.590776638
#> 71  071  1.311741930  2.023213268
#> 72  072 -2.463506961 -1.603139244
#> 73  073  0.756903502 -0.763596430
#> 74  074  0.000962237  0.112899042
#> 75  075 -0.374895370  0.075785394
#> 76  076  0.761584393 -0.027742678
#> 77  077 -0.792391608 -0.004328508
#> 78  078  0.567421380 -0.335833875
#> 79  079 -0.447336970  1.150941453
#> 80  080 -0.147475703  0.492689703
#> 81  081 -0.623711997 -1.598371931
#> 82  082  0.534833653 -0.573801750
#> 83  083 -0.889076757 -0.440083869
#> 84  084  0.555090385  1.459931363
#> 85  085 -1.009890038 -1.253896746
#> 86  086  1.288908008  0.313934485
#> 87  087 -0.006899855  0.166480553
#> 88  088  0.626442860  1.386717056
#> 89  089 -0.159931751 -0.245005066
#> 90  090  0.396582128  0.204678794
#> 91  091 -0.757759161  0.780409574
#> 92  092 -0.045529394 -0.495285935
#> 93  093 -1.071673082 -1.000797467
#> 94  094  0.553747041  0.309766011
#> 95  095  0.595871163 -0.106877625
#> 96  096 -0.649506146 -0.280863493
#> 97  097 -1.020394075 -1.297224987
#> 98  098  0.251779195 -0.792597382
#> 99  099 -0.351289679 -0.014957289
#> 100 100  0.208493122  0.633988002

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.71571023  0.882583862
#> 2   002 -0.97540136 -1.258861538
#> 3   003 -1.06238028 -0.542540446
#> 4   004  1.98826650  1.003141707
#> 5   005  1.98526955 -1.656004058
#> 6   006  0.20692007  0.678168089
#> 7   007  1.87459505  0.731580169
#> 8   008  0.23421351 -0.221223108
#> 9   009  0.36959377 -0.333838386
#> 10  010 -0.01779453 -1.326725961
#> 11  011 -0.81023516  0.255409113
#> 12  012 -0.25962163 -1.211509881
#> 13  013 -0.49959796  0.400474294
#> 14  014  0.04653671 -0.882729299
#> 15  015 -1.22355300 -0.343322925
#> 16  016  1.63587779  0.092870919
#> 17  017 -0.39872722  0.118386636
#> 18  018 -1.95815929 -1.593488410
#> 19  019 -1.12781306 -1.025015107
#> 20  020  0.88028413 -0.490739245
#> 21  021 -0.17569382  1.360139167
#> 22  022 -0.81482027 -0.999743229
#> 23  023  1.49778279  0.500111083
#> 24  024 -1.13846993 -0.488935247
#> 25  025 -0.33513898 -0.144394141
#> 26  026 -1.90193586 -0.378778137
#> 27  027 -1.49957999 -1.568454859
#> 28  028 -1.10764477  0.234728760
#> 29  029  0.06658475 -0.429681873
#> 30  030  0.09634989  2.136660210
#> 31  031 -0.06902713 -0.843400617
#> 32  032 -0.77856208 -1.523967949
#> 33  033  1.99506893  0.134101943
#> 34  034 -0.23900584 -0.673907663
#> 35  035 -0.88299353 -1.523457167
#> 36  036  0.28939231 -0.352010856
#> 37  037 -0.84832502 -0.991386541
#> 38  038  1.02310426  1.358566830
#> 39  039 -0.18788802  1.469386053
#> 40  040 -2.34036991 -1.808554286
#> 41  041 -2.69983918 -1.276383626
#> 42  042 -0.10843562 -0.976389926
#> 43  043  0.73223482 -0.220686448
#> 44  044  0.92365732 -0.278162455
#> 45  045 -0.43618107 -0.607899701
#> 46  046  1.35446656  1.197773183
#> 47  047  1.56210473  0.078174172
#> 48  048  0.50432412  2.023726822
#> 49  049 -0.17802423  0.594749502
#> 50  050 -1.23764997  0.113046709
#> 51  051 -0.95945344  0.172594064
#> 52  052  1.52933919  1.107616071
#> 53  053  0.76541559  0.379774188
#> 54  054  0.64469613  0.765230229
#> 55  055  0.86184114 -0.210739233
#> 56  056 -1.08410511 -1.527507218
#> 57  057  0.36167394  0.010852444
#> 58  058  0.77710129  1.075283546
#> 59  059 -0.25244015 -0.567582626
#> 60  060 -0.12377506  2.339905987
#> 61  061  1.56615421  0.244092382
#> 62  062 -0.84298203 -0.903783706
#> 63  063 -0.78774138  0.514017666
#> 64  064 -0.13515931  0.750808938
#> 65  065 -0.51926330  0.874292763
#> 66  066  0.15180762  0.618412554
#> 67  067  1.17784749 -0.865152277
#> 68  068 -0.26295774  0.050100211
#> 69  069  0.16668610 -0.816139080
#> 70  070  1.11982001  0.000551545
#> 71  071 -0.22256971 -0.363087050
#> 72  072  1.10292638 -0.645733080
#> 73  073  0.10571342  0.769284230
#> 74  074 -0.39683418  0.571089579
#> 75  075 -0.60735099  0.288767402
#> 76  076 -0.09955752 -1.051367528
#> 77  077 -1.24674379 -0.590243677
#> 78  078 -0.16452597  1.199708658
#> 79  079 -0.10734138 -1.813606812
#> 80  080  0.81001670  0.604685548
#> 81  081 -2.30618160 -0.513827552
#> 82  082 -0.58439568 -2.145060873
#> 83  083  1.39365483  1.893190864
#> 84  084 -0.04240583  1.852385787
#> 85  085  0.49305697  0.747447542
#> 86  086 -1.70658667 -2.492227044
#> 87  087  1.22763738  1.352263727
#> 88  088 -0.75213113 -1.149209808
#> 89  089 -1.96854778  0.426915479
#> 90  090  1.47669476  0.611155191
#> 91  091 -2.35175773 -2.381646619
#> 92  092 -0.16340172  0.748945549
#> 93  093  1.66840683  0.351755469
#> 94  094  0.42639727  2.050882737
#> 95  095 -1.39580436 -0.709126634
#> 96  096  0.78338246  1.857288412
#> 97  097  0.70051735 -0.204849070
#> 98  098  0.70698774 -0.552859367
#> 99  099 -0.12432380  0.120079224
#> 100 100 -0.44904157 -0.910076591