Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5025001
#> Y_2 0.5025001 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -1.83379518 -0.4332356
#> 2 -0.08780618 -0.7412878
#> 3 -0.40510525 -0.5768232
#> 4 -0.52916603  0.2584628
#> 5  0.52285274  0.9290244

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.60459856  1.51859949
#> 2   002  0.10076941 -0.46194344
#> 3   003 -1.02936772 -1.00327643
#> 4   004 -0.91365174 -0.24313170
#> 5   005  1.97873017  0.50833300
#> 6   006  0.16930618  0.86606899
#> 7   007 -1.49008304  0.47406090
#> 8   008 -0.45154941 -3.10180897
#> 9   009  0.71877056  0.61833180
#> 10  010 -0.63472007 -0.78147061
#> 11  011 -1.75431553 -0.98549304
#> 12  012  0.04842365  0.54719571
#> 13  013 -0.37337753 -0.76671632
#> 14  014  1.05782858  1.33478391
#> 15  015  1.15551603 -0.12400717
#> 16  016  1.14009670  0.21786737
#> 17  017  0.47893618  0.33546507
#> 18  018 -0.60207517 -0.09017029
#> 19  019 -1.42771710 -2.80055803
#> 20  020 -0.40627781 -1.56417053
#> 21  021 -0.11972847 -1.03720513
#> 22  022 -0.04443309 -0.91144573
#> 23  023 -0.43746250  0.14933417
#> 24  024  0.56717920  0.69367023
#> 25  025  1.31469495  0.45152149
#> 26  026 -0.73515694 -2.10635315
#> 27  027  0.70345314 -0.34821935
#> 28  028 -2.34158345 -1.05102973
#> 29  029 -0.63606999 -0.18820998
#> 30  030  0.46256068 -1.14949680
#> 31  031 -1.70681980 -0.57151066
#> 32  032  0.03742736  0.54837976
#> 33  033  0.47329675  0.38547848
#> 34  034  0.58922349 -0.92675611
#> 35  035  0.76991954  0.74868324
#> 36  036 -0.36077100 -1.24308604
#> 37  037  1.94377363  0.75837545
#> 38  038 -0.14293523  1.41500049
#> 39  039 -1.07827654  1.13977358
#> 40  040 -1.79126157 -1.69786991
#> 41  041  0.15224487 -0.77069419
#> 42  042 -2.21153992 -0.05737732
#> 43  043  0.86756626  1.03512347
#> 44  044 -1.38409125 -0.99165115
#> 45  045 -0.30755010  0.43175498
#> 46  046  1.90848359  0.02661270
#> 47  047 -0.28230816  0.37784607
#> 48  048 -0.20135383  1.35909478
#> 49  049 -0.94527004 -1.16765428
#> 50  050 -0.96415854 -0.20159003
#> 51  051  0.99219402  0.84921593
#> 52  052 -0.14758595  1.13903849
#> 53  053 -0.66324637 -0.37865428
#> 54  054 -0.60038056 -0.71211818
#> 55  055  0.08235235  0.17445080
#> 56  056  0.10729621 -0.26047243
#> 57  057  0.27233217 -0.87313216
#> 58  058 -1.64039532 -0.39689388
#> 59  059  0.74380797 -0.34446645
#> 60  060  0.17244982 -0.31797541
#> 61  061  0.41249558  0.38412444
#> 62  062  0.43962300  0.36446725
#> 63  063  1.55196344  1.89481306
#> 64  064 -0.03011047 -1.92389493
#> 65  065 -0.25200313 -1.00608540
#> 66  066  1.46402230 -0.33551118
#> 67  067 -0.44199140 -0.43003171
#> 68  068 -0.82151887 -1.01959234
#> 69  069 -1.56483268 -1.47159513
#> 70  070  1.04934586 -0.19703219
#> 71  071 -2.16106931 -1.52540757
#> 72  072  0.04614158  0.64730766
#> 73  073 -1.16816689 -1.15096076
#> 74  074  0.57538620  0.12869050
#> 75  075 -0.94777736 -0.80268782
#> 76  076 -0.30816333  0.09174424
#> 77  077 -0.51817253  0.56306972
#> 78  078  1.66816883  0.55581785
#> 79  079 -0.15489051 -1.00460784
#> 80  080  1.22319403  0.35294698
#> 81  081  0.91035025  1.28922920
#> 82  082  0.09070855  0.31328685
#> 83  083  0.96361964  1.14488099
#> 84  084 -1.65920819 -1.94945251
#> 85  085 -0.39726768 -0.49919588
#> 86  086 -0.32959934  0.35441236
#> 87  087  0.64663927  0.57489909
#> 88  088  0.31901751  0.43657274
#> 89  089  0.36871932  1.40402325
#> 90  090  1.65421466  1.96977977
#> 91  091 -1.35937436 -0.31140259
#> 92  092 -0.77569897  0.05162577
#> 93  093 -1.23805701 -1.53237600
#> 94  094  1.15616283  1.44510207
#> 95  095 -1.05867866 -0.43727454
#> 96  096  1.02573568  1.10031395
#> 97  097 -1.19438470 -1.14126365
#> 98  098  1.00568046  0.61867504
#> 99  099  0.20897207  1.43584975
#> 100 100  1.26624278  1.11769175

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.16502641  0.058836446
#> 2   002 -0.35720027  0.418926007
#> 3   003 -0.11432990  0.356530515
#> 4   004  1.54180836  0.514225741
#> 5   005 -0.91030597  0.302338717
#> 6   006 -1.15530206 -0.542282992
#> 7   007  0.77836212  0.390906752
#> 8   008  0.49485895  0.610776263
#> 9   009 -1.60636969 -1.024632395
#> 10  010  0.37963359  0.473590774
#> 11  011 -1.01765426 -1.048346294
#> 12  012  0.21595728  0.551648006
#> 13  013 -0.06594970  1.249049555
#> 14  014 -1.61707723 -0.493893819
#> 15  015  1.22109084  1.396853516
#> 16  016 -2.30839612 -0.716588190
#> 17  017  0.83358558  0.779781165
#> 18  018  1.95077107  1.836541373
#> 19  019 -1.77703895 -0.312597263
#> 20  020 -1.23665063 -1.418553911
#> 21  021  0.16225935  0.569905426
#> 22  022  0.19619926  0.736582938
#> 23  023 -1.19202197 -1.257453754
#> 24  024 -0.19421975 -0.441517124
#> 25  025 -2.23136742 -0.476553851
#> 26  026  1.70123140  2.048388783
#> 27  027  0.32294945  1.834297102
#> 28  028  0.96779981 -0.017664819
#> 29  029  0.04951448 -0.258722961
#> 30  030  0.03483588  0.624417198
#> 31  031 -2.07731702 -1.300610744
#> 32  032  1.41303272 -0.002286861
#> 33  033  0.79605992  0.674173989
#> 34  034 -1.50659779 -0.212808305
#> 35  035  1.09969972  0.653701546
#> 36  036 -1.73470117 -0.408707614
#> 37  037 -1.08811050  0.328038641
#> 38  038  0.61265502  0.399934107
#> 39  039 -0.07764898  0.596477853
#> 40  040 -0.68152828  0.528469710
#> 41  041  0.93021993  0.198477558
#> 42  042  0.76126167  0.880195719
#> 43  043 -0.84843906 -1.006316080
#> 44  044  0.19941807  0.184146639
#> 45  045  1.31840567  2.172352214
#> 46  046  0.23212482 -0.176947095
#> 47  047 -0.96625856 -2.276677541
#> 48  048 -0.57872972 -0.100335012
#> 49  049 -0.20689705  0.807597362
#> 50  050  0.28619996 -0.790900650
#> 51  051 -0.17201130  0.561901514
#> 52  052  0.83452843  0.666873762
#> 53  053  0.74194387  0.130868988
#> 54  054  0.41801731 -0.107295062
#> 55  055  1.56609378  2.033242682
#> 56  056 -0.48518889 -0.590939678
#> 57  057 -0.69288156  0.368659963
#> 58  058 -0.98183889  0.096507967
#> 59  059  0.49441987  1.092998930
#> 60  060 -0.42722844 -1.025813711
#> 61  061  1.37863082  0.935277775
#> 62  062 -0.70767352 -1.178386444
#> 63  063  0.21744927 -0.685180659
#> 64  064 -2.16212968 -0.951747442
#> 65  065 -0.22638331  0.417369360
#> 66  066  0.26727174  0.712842467
#> 67  067  0.30774006  1.234490232
#> 68  068 -0.01626032  0.150852789
#> 69  069 -0.28694180 -1.510562374
#> 70  070 -2.07957549 -1.064506593
#> 71  071 -0.56530001 -1.186920088
#> 72  072 -0.41609208  0.487419259
#> 73  073  0.66115813  0.350528772
#> 74  074 -0.57259025  1.130789159
#> 75  075 -1.27520644 -0.897516231
#> 76  076 -0.15898513  0.232588786
#> 77  077 -0.17157890  0.991057791
#> 78  078  2.25957086  1.477711518
#> 79  079 -0.26745977 -1.648963305
#> 80  080 -0.64458133 -1.244525047
#> 81  081 -0.55691569  0.454773464
#> 82  082  0.85831247  0.404855180
#> 83  083 -0.39596513 -0.341679704
#> 84  084  0.08060543 -0.637985538
#> 85  085 -0.40762308 -1.000538118
#> 86  086  0.18369993 -0.234279421
#> 87  087 -1.66912005 -1.504396714
#> 88  088 -0.67586574 -0.169709245
#> 89  089  0.63048034 -0.720983944
#> 90  090  1.34095029  0.748641519
#> 91  091 -0.44204035  0.442050370
#> 92  092 -1.05789874  0.164343337
#> 93  093 -0.29513476  0.126086676
#> 94  094 -1.16093792  1.143182693
#> 95  095 -0.40878644  0.199902794
#> 96  096  1.96371575  0.494818044
#> 97  097 -1.24250050  0.154437632
#> 98  098 -0.71572811  0.542878649
#> 99  099  0.43462460 -0.416866085
#> 100 100 -0.02407118 -0.142134775