Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5159341
#> Y_2 0.5159341 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1         Y_2
#> 1  2.759805  1.68650386
#> 2  1.460459  0.09617979
#> 3  1.164937 -0.74748402
#> 4  1.315834  0.31709640
#> 5 -1.453248 -0.46623494

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.62237385 -0.99572092
#> 2   002 -0.14162274  1.59491937
#> 3   003  1.56212192  0.78204916
#> 4   004  1.64540627  1.94213669
#> 5   005 -0.75333402 -0.78624808
#> 6   006 -0.66546361 -0.73506478
#> 7   007  0.89599127  0.16786792
#> 8   008 -1.51196325 -0.96238488
#> 9   009 -0.98780953 -0.77777911
#> 10  010  0.07593694 -0.07940524
#> 11  011 -1.31052739 -0.02357398
#> 12  012  0.06846716  0.09814194
#> 13  013 -0.94954102 -0.61374100
#> 14  014 -0.71019293 -0.49457443
#> 15  015 -1.21773524 -0.64231460
#> 16  016 -0.45471278 -0.83759572
#> 17  017  0.54758501  0.78408262
#> 18  018  0.82700291 -0.08815427
#> 19  019  0.64172838 -0.91353394
#> 20  020  0.85825362 -1.11848929
#> 21  021 -0.06772640 -0.22637767
#> 22  022 -0.67221869 -0.24121841
#> 23  023 -1.16354179 -0.37264914
#> 24  024  1.53311529  1.62042597
#> 25  025  3.27650311  1.94797752
#> 26  026  0.70361250  0.23413011
#> 27  027  1.58463568 -0.02539047
#> 28  028 -0.75434266 -0.09271476
#> 29  029  0.18812543 -0.30842413
#> 30  030  0.30024559  0.64923825
#> 31  031  0.83878949 -0.55136740
#> 32  032  0.39752418  0.40669017
#> 33  033 -0.96694942 -0.36921343
#> 34  034  0.66133025  1.00160296
#> 35  035  1.05315666  0.48456135
#> 36  036 -0.24742422 -0.07170625
#> 37  037 -1.41900949 -2.85970930
#> 38  038 -0.10921579  0.17949586
#> 39  039  0.19117301 -1.29740280
#> 40  040  1.31199171  1.19675468
#> 41  041  0.26176686 -0.02169928
#> 42  042  0.63592486 -1.12683406
#> 43  043 -0.56838782 -0.32931522
#> 44  044  0.25726288 -0.61772068
#> 45  045  0.95102036 -1.20631844
#> 46  046 -0.35196921  0.34358959
#> 47  047  1.20576993  1.28404080
#> 48  048 -1.05508169  0.11085644
#> 49  049 -0.15383599 -0.17619278
#> 50  050 -0.68922710 -0.99654630
#> 51  051  0.09362867 -1.41621310
#> 52  052 -0.10009650  0.31604373
#> 53  053 -1.23901447 -0.65069933
#> 54  054 -0.70180707 -0.13815862
#> 55  055 -0.54222248 -0.24352064
#> 56  056  0.52944837  0.63739271
#> 57  057 -0.12090193  0.90264709
#> 58  058 -1.03620969  0.16447216
#> 59  059  1.40534345 -0.32645165
#> 60  060  1.45025004  2.08729791
#> 61  061  0.76404224  0.50293663
#> 62  062  0.18814148  2.21102913
#> 63  063  1.30196688  0.66399420
#> 64  064 -0.26229727 -0.43322759
#> 65  065 -1.33047798 -0.21835374
#> 66  066 -0.51470662 -0.28437037
#> 67  067 -0.51293630 -2.11913834
#> 68  068  1.04688238  1.21069215
#> 69  069 -0.32928060 -0.08803660
#> 70  070  0.75554147  0.37432367
#> 71  071  0.93616901  1.35514138
#> 72  072  0.96945332  0.94589010
#> 73  073 -0.32248171 -0.71664607
#> 74  074  1.90620999  1.19685986
#> 75  075 -0.80235250 -0.97311983
#> 76  076  0.15183936  0.42909517
#> 77  077  0.28771314 -1.13351648
#> 78  078  0.26794258  0.66069069
#> 79  079  0.49230404  0.41954093
#> 80  080  1.72964573  1.51977281
#> 81  081 -0.99598593  0.57692706
#> 82  082 -0.64341345  0.40499151
#> 83  083  0.75961576  0.06629028
#> 84  084 -0.58079275 -1.68666261
#> 85  085 -1.38752984 -1.60115879
#> 86  086  0.33980676  1.10174456
#> 87  087 -0.20285420 -0.23596287
#> 88  088 -0.02641299  1.48154556
#> 89  089 -0.49191710 -0.47887223
#> 90  090  0.69797136 -0.22999440
#> 91  091  0.24913642  0.60824081
#> 92  092 -1.63013397 -0.71452443
#> 93  093  0.72014858  0.55511229
#> 94  094 -2.30833970 -1.39042114
#> 95  095  0.88042490  1.14859192
#> 96  096 -1.45177506 -0.80470745
#> 97  097 -0.36418676  0.52457336
#> 98  098 -2.50885839 -2.24814960
#> 99  099  0.19643969 -0.23983119
#> 100 100  0.53506627 -1.50277019

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.62097326  0.447826871
#> 2   002  0.84033148 -0.599600361
#> 3   003  0.25013831  1.145383144
#> 4   004  0.49516632  1.795735068
#> 5   005  1.01188731 -0.176781318
#> 6   006  1.12469321  0.123182885
#> 7   007  0.97506664  1.203376857
#> 8   008  2.07105066  0.093328514
#> 9   009 -0.71204534 -1.891889284
#> 10  010  1.48530291  1.066966373
#> 11  011  0.79401109  0.517287105
#> 12  012 -0.30871156 -0.102964713
#> 13  013  0.94708767  0.710232088
#> 14  014  0.14141152 -0.473836419
#> 15  015 -0.34887024  1.262098407
#> 16  016  0.70526845 -0.317361029
#> 17  017  0.06156753  0.546660848
#> 18  018  0.13645885 -0.349978248
#> 19  019  0.23509880  0.533270498
#> 20  020 -1.83938125  1.374827815
#> 21  021 -1.51010967 -0.633656267
#> 22  022  0.17642468  0.875188079
#> 23  023  0.12886898  0.352939788
#> 24  024  0.83582090  1.015962488
#> 25  025 -0.45917116 -0.549665923
#> 26  026 -1.19026346 -0.884885851
#> 27  027 -0.76158919  0.535039852
#> 28  028  1.65509272  0.194342208
#> 29  029  0.21509045  0.203019067
#> 30  030 -0.49677710 -1.242381081
#> 31  031 -0.23066632  0.358019312
#> 32  032  1.42710321 -0.173405061
#> 33  033 -0.48475872 -1.660433733
#> 34  034 -0.67813429 -1.862858319
#> 35  035  0.23499636 -1.306597604
#> 36  036 -1.38941833 -1.151477867
#> 37  037 -0.07075454  1.249690346
#> 38  038 -0.10894354  0.309508482
#> 39  039  0.87293920  1.515014715
#> 40  040  0.83712379  0.025609567
#> 41  041 -1.30636232 -0.254302043
#> 42  042  1.51922434  1.724559734
#> 43  043 -0.05539073 -0.257501446
#> 44  044 -0.15278098  0.069070794
#> 45  045 -0.31150598 -0.072240308
#> 46  046  0.27430066  0.482754276
#> 47  047 -0.25760642 -0.608845719
#> 48  048  1.13583044  0.026122437
#> 49  049  0.54505761 -0.006148566
#> 50  050  0.40824822  1.032019895
#> 51  051 -2.03004173 -0.104197790
#> 52  052 -0.22590580 -2.373342193
#> 53  053 -1.26137160 -0.625264263
#> 54  054 -1.34866022 -0.298144983
#> 55  055 -1.78255118 -1.172779569
#> 56  056 -0.27969085 -0.139337783
#> 57  057 -0.45030719 -0.258508822
#> 58  058  0.17107712 -0.117861893
#> 59  059 -0.49886851 -1.403668469
#> 60  060  2.05084419 -0.246691718
#> 61  061  1.05171886  0.153789013
#> 62  062  0.34411964  0.811766932
#> 63  063 -0.88393862  2.104363621
#> 64  064 -0.25186520  0.268771253
#> 65  065  0.07110288 -1.010680009
#> 66  066 -0.50855458 -0.882257395
#> 67  067 -0.43112472 -0.073455078
#> 68  068 -0.33835385  0.233216640
#> 69  069 -0.14902694 -0.361626666
#> 70  070 -0.52266994 -0.792556356
#> 71  071  0.42475338 -0.824779336
#> 72  072 -0.16606419  0.269958216
#> 73  073  0.48587566  0.236406234
#> 74  074  0.33685170 -0.810295889
#> 75  075  0.43638720  1.065138549
#> 76  076 -1.29530123 -0.218797816
#> 77  077 -0.63929667 -2.110966766
#> 78  078  0.25693226  0.680408385
#> 79  079 -1.41674188 -0.898037371
#> 80  080  0.94848495  1.098412972
#> 81  081 -0.91667017 -0.126310415
#> 82  082 -0.22510325 -0.926386157
#> 83  083 -1.28230707 -2.029573782
#> 84  084  0.67085297  1.548847515
#> 85  085 -1.71970454  0.709558265
#> 86  086 -0.33589698 -0.740649897
#> 87  087  0.01321849 -0.281390838
#> 88  088 -1.53467261  0.880767398
#> 89  089  0.33935845 -1.282239403
#> 90  090 -0.49626884 -0.433084267
#> 91  091  1.74013967  1.304064481
#> 92  092 -0.93285802 -0.099842017
#> 93  093 -0.08765187  0.692516665
#> 94  094 -0.36687904 -1.300686721
#> 95  095 -1.60480436 -0.728757361
#> 96  096  2.58086435  1.953901144
#> 97  097 -0.34406511  0.857519675
#> 98  098  0.71263256 -0.163126156
#> 99  099  0.32991048 -0.451499159
#> 100 100  0.11411217 -1.901000883