Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5006171
#> Y_2 0.5006171 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.2229554 -0.5926935
#> 2 -0.0726107  0.7167665
#> 3 -1.3185606 -0.7709426
#> 4  1.9245557  0.8650085
#> 5  0.7846748  1.4697404

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.06875078  1.327966884
#> 2   002  0.34977983  0.274294498
#> 3   003 -0.52224101 -0.626343307
#> 4   004 -0.98630177 -0.691742132
#> 5   005  0.03889939  0.161484006
#> 6   006 -1.40477582 -0.696822619
#> 7   007 -0.26963868 -0.499149913
#> 8   008  0.06251672  0.627065547
#> 9   009  0.48890923 -0.098531076
#> 10  010  0.06682343  1.779865543
#> 11  011  1.53347138  2.837669859
#> 12  012 -0.65839168 -1.456161909
#> 13  013  1.33618110  0.751323768
#> 14  014  0.83516285  0.375968526
#> 15  015 -1.05153468 -0.448276125
#> 16  016  0.61077283  1.582478343
#> 17  017 -1.29242712 -0.302840773
#> 18  018 -1.95185364 -1.118887675
#> 19  019  0.17328702  1.239106586
#> 20  020  0.35856088  0.858643555
#> 21  021 -1.95001176 -1.350009810
#> 22  022  1.02399442  2.497398507
#> 23  023  0.20120856  0.573095960
#> 24  024  1.45487110  1.134365569
#> 25  025  0.90448252 -0.561193963
#> 26  026 -0.21138223 -0.265258986
#> 27  027 -0.38049484 -0.489418189
#> 28  028  1.56824415  1.702792909
#> 29  029 -0.74474159 -1.001435369
#> 30  030  1.41592658 -0.804890715
#> 31  031  0.29103366  2.058057108
#> 32  032 -0.58744933 -0.735820978
#> 33  033 -0.45667459 -0.712194195
#> 34  034 -0.29547887  1.308730922
#> 35  035 -1.08718147 -0.852432742
#> 36  036  0.38235292 -1.734808219
#> 37  037 -0.64405503 -0.446746840
#> 38  038 -1.66609537 -0.990782217
#> 39  039  1.29162900  1.621052069
#> 40  040 -0.67239299 -0.309560097
#> 41  041 -0.14572010 -0.665749739
#> 42  042  0.75315606  0.933208572
#> 43  043  1.32500699  0.743451070
#> 44  044 -1.39445081  0.209581581
#> 45  045 -2.14904621 -0.233280189
#> 46  046 -1.86656818 -1.653677797
#> 47  047 -1.01524873 -0.284089916
#> 48  048 -0.45896095  0.908426454
#> 49  049 -1.40598736 -0.773399882
#> 50  050 -0.60950157 -1.091296850
#> 51  051 -1.14991926 -0.074578619
#> 52  052 -1.07374548  0.030772317
#> 53  053  0.95619514  0.020676660
#> 54  054  1.67871378  0.922592138
#> 55  055  0.64302005  0.197611506
#> 56  056  0.07912180 -0.236234957
#> 57  057 -0.18040946 -0.108799185
#> 58  058 -0.33865395  0.575604247
#> 59  059  1.99363622 -0.001052688
#> 60  060 -0.61920862  0.428496850
#> 61  061 -0.98967583 -0.443246005
#> 62  062  0.66998288  0.428496180
#> 63  063  0.99236760  0.387256392
#> 64  064 -0.53552165 -0.249273879
#> 65  065 -0.18369211  0.305120301
#> 66  066  1.66834739  2.845663411
#> 67  067  1.42401593  0.801424308
#> 68  068 -1.12124960 -0.375972662
#> 69  069  1.53088861  1.260398548
#> 70  070  0.66455903  0.782698846
#> 71  071  1.43175872  0.658539885
#> 72  072 -0.94332866 -0.971428585
#> 73  073 -0.01617173  1.038760564
#> 74  074  0.43896607  0.045921129
#> 75  075 -0.70500511 -0.142058098
#> 76  076 -1.11836917 -0.913610647
#> 77  077  0.31347136  0.803939115
#> 78  078  0.13985962 -0.675745395
#> 79  079 -0.18212581 -0.706886310
#> 80  080  2.20710391  0.938597637
#> 81  081 -0.08550870  0.684054459
#> 82  082  2.06379454  1.408387002
#> 83  083  0.72349824 -0.990950133
#> 84  084  0.39244338  0.492644641
#> 85  085 -0.92383337 -0.515974641
#> 86  086  1.40545585  0.914726142
#> 87  087  0.25240784  0.521116034
#> 88  088 -0.50294971 -0.104182381
#> 89  089  0.77319900  0.877954743
#> 90  090  2.16350248  1.040975866
#> 91  091  0.01625648  0.629695219
#> 92  092 -0.25120074  1.391804193
#> 93  093 -2.53378815 -2.212249927
#> 94  094  0.70230999 -0.291133257
#> 95  095 -0.26938528 -0.927744082
#> 96  096 -0.21237516  0.757850255
#> 97  097 -1.64139254  0.176621940
#> 98  098 -0.25378853  1.736627480
#> 99  099  0.67737768  0.026276370
#> 100 100 -0.18852900  2.100223866

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  3.121895821  2.415053355
#> 2   002  0.332518563 -0.110936135
#> 3   003  0.096247407  0.455517097
#> 4   004  0.408168972 -0.059440470
#> 5   005  1.917260048  2.075090488
#> 6   006 -0.887915641 -0.164885101
#> 7   007 -1.811009616 -2.149144517
#> 8   008 -0.092312627  1.177185309
#> 9   009  2.393146110  2.203911934
#> 10  010 -0.175495269 -0.123910187
#> 11  011  1.349436261  1.202527523
#> 12  012  0.545384784  1.393396492
#> 13  013  1.121125532  0.100945061
#> 14  014  1.364372734  0.501888719
#> 15  015  0.568346815  1.384871659
#> 16  016  0.636323589 -1.261189214
#> 17  017  0.767926899  1.773300011
#> 18  018  0.849912367  1.112594557
#> 19  019  0.080793513  0.275130731
#> 20  020  0.677050711  1.200671198
#> 21  021 -0.318157320  0.770568493
#> 22  022 -1.150216701 -1.711856918
#> 23  023 -0.099122633 -0.512924095
#> 24  024 -0.466620902  0.724425747
#> 25  025 -0.651352124  0.093387656
#> 26  026 -0.033913335  0.226652804
#> 27  027 -0.720696037  1.118893407
#> 28  028  0.370534353  1.277934345
#> 29  029  1.180149611  0.875510626
#> 30  030 -0.027570084 -0.131177309
#> 31  031  1.383819965  2.093345368
#> 32  032 -1.338611287  0.244076115
#> 33  033 -0.239881385  0.607343278
#> 34  034  1.311414797  0.515843784
#> 35  035  0.656896057  2.560826798
#> 36  036 -1.854732324  0.849644000
#> 37  037  1.700975791  0.666658810
#> 38  038 -1.649327252 -1.624458942
#> 39  039  0.193574434  0.004323576
#> 40  040 -0.776725632 -0.534525782
#> 41  041 -0.070071760  0.871045682
#> 42  042 -1.059432876  0.523645473
#> 43  043  2.574915530  3.137415627
#> 44  044 -0.963339893  0.399554321
#> 45  045  0.426871131  0.944937840
#> 46  046  1.444446180 -0.184355789
#> 47  047 -1.023647086  0.041583471
#> 48  048  0.287889709  0.106228291
#> 49  049 -1.354084225 -0.678103947
#> 50  050  1.098877033  1.688156961
#> 51  051 -0.665106442  1.099132511
#> 52  052  0.817829277 -0.625220349
#> 53  053 -0.056586791  1.159913016
#> 54  054 -0.518578014  0.006253672
#> 55  055 -0.122290644 -0.384565646
#> 56  056 -0.780642478 -1.039143079
#> 57  057  1.310682894  0.570702654
#> 58  058 -1.223593217 -0.290367907
#> 59  059  1.331704977  0.514394655
#> 60  060 -0.494554419  0.752924008
#> 61  061 -0.424842512  1.750462882
#> 62  062 -0.336623457  0.347047083
#> 63  063 -0.619711886  0.545253282
#> 64  064 -0.823832764  0.593607363
#> 65  065  0.970325245  0.577146094
#> 66  066 -0.325000007  1.188748514
#> 67  067  0.002837736  1.234631189
#> 68  068 -1.356497512  0.126301677
#> 69  069 -0.119817103  1.315311631
#> 70  070 -0.293870132 -0.695792093
#> 71  071 -0.615439635  1.579194411
#> 72  072  1.264753059  2.519511985
#> 73  073 -0.240442056 -0.595481490
#> 74  074  0.354737795 -0.497889639
#> 75  075 -0.754623502  0.571352900
#> 76  076 -0.848136353 -1.441382024
#> 77  077  2.724748859  1.875986361
#> 78  078  0.439836613  1.453930953
#> 79  079 -1.296854904  0.256664844
#> 80  080 -0.773745481 -0.367304857
#> 81  081 -0.496110347  0.187561044
#> 82  082  0.469330875  0.728655301
#> 83  083  0.480577598  1.245164843
#> 84  084  0.234699786 -1.618606692
#> 85  085  0.846076647 -0.603866622
#> 86  086  2.103240265  1.060680272
#> 87  087  0.707386150  1.032989245
#> 88  088  2.532161420  0.963478392
#> 89  089  1.453572228  1.762601768
#> 90  090  1.587136062  1.218372495
#> 91  091  0.836494870  1.705107129
#> 92  092 -0.537130963 -0.803270916
#> 93  093 -0.605695094  0.328092938
#> 94  094  1.459988294 -0.085671793
#> 95  095  0.022985679  0.095293341
#> 96  096  1.271524280  1.503276152
#> 97  097 -0.057890393  0.889166953
#> 98  098 -0.188193735  0.484431715
#> 99  099  1.289829297  0.589503670
#> 100 100  0.362589163 -1.250040978