Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4995507
#> Y_2 0.4995507 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.9311810  0.6177901
#> 2  0.6391573  0.5324074
#> 3  1.2709136  1.6573133
#> 4  2.4335256 -0.1965581
#> 5  0.2135019  0.5192013

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.299460082 -1.83966986
#> 2   002  0.047827526 -0.59902518
#> 3   003  0.089615133  1.20905762
#> 4   004  1.514218224  0.49166754
#> 5   005  1.634533754  0.64953241
#> 6   006  0.216883397  1.40710957
#> 7   007 -1.583490324 -0.57934405
#> 8   008  0.999508944  0.78900446
#> 9   009  0.416043296  0.50436449
#> 10  010  0.918773505  1.31723850
#> 11  011  1.354746702  0.76227255
#> 12  012  0.437648926  0.33273272
#> 13  013 -0.432769377 -0.51387072
#> 14  014  0.472865931 -0.37539927
#> 15  015 -0.916893588 -0.62752705
#> 16  016 -0.282105126  0.94061701
#> 17  017 -1.716382072 -0.82078400
#> 18  018  0.294064969  0.09950882
#> 19  019 -1.627646301 -1.81957402
#> 20  020 -0.947061956 -0.94081428
#> 21  021 -0.183281230  0.36755785
#> 22  022  1.680779843  1.79182324
#> 23  023 -0.930816811  0.38621065
#> 24  024  0.495049851  0.01359894
#> 25  025  2.248633696  0.85126364
#> 26  026 -0.945662827  0.26459672
#> 27  027  0.162258588 -0.89375933
#> 28  028  0.098391538  0.80493051
#> 29  029  0.773463201  1.68826630
#> 30  030 -0.580887101  0.06323389
#> 31  031 -0.619240967 -0.26536573
#> 32  032 -0.783100375 -0.91316859
#> 33  033  0.077790023  0.87714016
#> 34  034  0.087298299  0.09472309
#> 35  035  0.854602555  0.19753709
#> 36  036  0.303838489  1.37370674
#> 37  037 -0.892692290  0.12780852
#> 38  038  0.738450459  1.11323629
#> 39  039 -1.691457231 -1.33748556
#> 40  040 -0.403015186  0.11686175
#> 41  041  0.830833595  0.11260464
#> 42  042 -0.966798351 -0.16990134
#> 43  043  0.239698763  0.31793132
#> 44  044  0.963155093 -0.46448217
#> 45  045  0.002705209  0.77502778
#> 46  046 -0.340164542  1.08797259
#> 47  047  0.160047188 -1.10752011
#> 48  048  0.639376807  0.51300701
#> 49  049 -0.717111067 -0.35766961
#> 50  050  2.297426113  0.25964996
#> 51  051  0.182703942  0.31141674
#> 52  052  0.124003047 -0.07865841
#> 53  053 -1.998565955 -1.06269317
#> 54  054  1.602482314  1.14676532
#> 55  055 -0.215991438  0.09968860
#> 56  056 -0.521304069 -1.03089670
#> 57  057 -1.761051727 -0.14371573
#> 58  058  0.820666491 -0.98347951
#> 59  059 -0.477233545  0.40504772
#> 60  060 -0.390158758  0.50555687
#> 61  061  2.052327153  1.42173992
#> 62  062  0.691363573  1.13120105
#> 63  063 -1.228023071 -1.25589003
#> 64  064  0.470149191  1.02549974
#> 65  065 -0.703387595 -0.37821486
#> 66  066  1.849786137  2.44584510
#> 67  067 -0.531911524 -0.53278457
#> 68  068  0.615334581  0.61684206
#> 69  069  0.774589665 -0.05381115
#> 70  070  2.148915062  1.30334215
#> 71  071 -0.545604629  0.12743334
#> 72  072 -0.286425947  0.09339732
#> 73  073  0.555256598  0.01328109
#> 74  074 -0.084892108  0.77025314
#> 75  075 -1.268945314 -0.22642409
#> 76  076 -0.224915927  0.81420094
#> 77  077 -0.229483365 -0.06983187
#> 78  078 -0.154876602 -1.07367760
#> 79  079  0.546480596  1.38820326
#> 80  080 -0.011687954  0.20178923
#> 81  081 -0.158713344 -0.08118028
#> 82  082  0.848231505 -0.35326197
#> 83  083  0.142455066  0.16389977
#> 84  084 -0.143442498 -1.22046044
#> 85  085  0.038995674  1.01312918
#> 86  086 -0.361157559 -1.66897764
#> 87  087  1.604045352  0.88026254
#> 88  088  0.685243012  1.07110931
#> 89  089 -1.026169947  0.97496110
#> 90  090  1.442456346  1.35557023
#> 91  091  1.839157682  0.34444285
#> 92  092  0.145749566 -0.48146621
#> 93  093  0.608707385  1.06698446
#> 94  094 -0.013246036  0.34992750
#> 95  095  0.697850706  0.10158207
#> 96  096 -0.566040294  0.35310712
#> 97  097  1.099240922  0.91679682
#> 98  098 -0.693453206 -0.48066273
#> 99  099 -0.997673181 -1.65908830
#> 100 100  1.526758110  1.42676501

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.87042507 -0.853945464
#> 2   002  1.05748173 -1.020012792
#> 3   003  0.31465148 -0.389534891
#> 4   004 -0.32143296 -0.441535884
#> 5   005  0.22457994  0.352733730
#> 6   006 -0.81496913 -1.025307228
#> 7   007 -0.76200479  0.342562830
#> 8   008 -0.81285278 -0.523639941
#> 9   009  0.84105544 -0.040605393
#> 10  010  0.02691526  0.101002693
#> 11  011 -0.72943568 -2.220306712
#> 12  012  2.20112582  1.537932787
#> 13  013 -1.43417849 -0.796844022
#> 14  014 -1.04057408 -0.996144928
#> 15  015  0.64278797 -1.525098900
#> 16  016 -0.80314503  0.647498980
#> 17  017 -0.72485636 -1.012285121
#> 18  018 -1.20885813 -0.459269207
#> 19  019 -1.75777093 -1.765702273
#> 20  020  1.26371724  0.983858121
#> 21  021  0.21393930 -0.284482453
#> 22  022 -1.91883255 -2.365643742
#> 23  023  0.18539866 -1.558569536
#> 24  024 -0.71013687 -0.069017268
#> 25  025  0.63054977 -0.296996338
#> 26  026  0.38241224  0.568251086
#> 27  027 -1.82958300 -0.472024411
#> 28  028 -2.31840235 -0.400075344
#> 29  029  0.55867993  1.647588652
#> 30  030  1.06019221  0.942411598
#> 31  031  0.95331867  2.328261286
#> 32  032 -0.86500034 -0.362732739
#> 33  033  0.18133453  0.424938044
#> 34  034 -0.66928274 -0.603714562
#> 35  035 -0.26863739  0.491669679
#> 36  036 -1.30550663 -0.060095361
#> 37  037  0.54485220 -1.119486854
#> 38  038  1.66209427  0.110992896
#> 39  039 -1.56375054 -1.537356494
#> 40  040 -0.64213941 -0.422533242
#> 41  041  0.74591468  0.283114612
#> 42  042  0.21747251  0.894248262
#> 43  043  0.57123881  0.343939606
#> 44  044  2.50712428 -0.206664976
#> 45  045  0.13793722 -0.148209055
#> 46  046  0.89793236 -0.248677492
#> 47  047 -1.20303823 -1.364644882
#> 48  048  0.91189497  0.533274125
#> 49  049 -1.67529556  0.905872083
#> 50  050 -0.26048834 -0.999864739
#> 51  051 -1.86217443 -1.278367951
#> 52  052 -0.90044466  0.079421955
#> 53  053 -2.92692485 -3.111253478
#> 54  054 -0.25515629  0.205508237
#> 55  055  0.05898262  0.195497468
#> 56  056 -0.47403506  1.267495057
#> 57  057  0.76148738  0.082289592
#> 58  058 -0.33835788 -1.941619719
#> 59  059 -0.99882799 -2.247347426
#> 60  060  1.01264926  0.575869576
#> 61  061 -2.67203821 -1.580862878
#> 62  062 -0.81144042 -0.440102919
#> 63  063  1.42318963  1.180385353
#> 64  064 -0.07921994 -0.737906364
#> 65  065 -1.37036202 -1.483215676
#> 66  066  0.66317654 -1.074068329
#> 67  067  0.12149776  0.115307487
#> 68  068  0.83978963  1.075835740
#> 69  069  0.87174597  0.578087556
#> 70  070  0.56795492 -0.138647038
#> 71  071 -1.53691501  0.017142979
#> 72  072 -0.32668039  0.416777630
#> 73  073 -0.46701928  0.306456963
#> 74  074 -0.70461262 -0.565242679
#> 75  075  0.77264809  1.562577558
#> 76  076 -0.09400787 -1.180207275
#> 77  077  0.19098115 -0.037768127
#> 78  078 -0.18397696 -0.199481611
#> 79  079  1.13705684  0.770900799
#> 80  080 -0.35954465 -0.411799983
#> 81  081  1.59916044  1.622682598
#> 82  082 -1.29576272 -0.135048810
#> 83  083 -0.73794971 -1.154508041
#> 84  084 -1.55105223 -1.048914800
#> 85  085 -0.86989485  0.313318575
#> 86  086 -0.86094283  0.891244755
#> 87  087  1.05461458 -0.933739351
#> 88  088 -1.44402441 -0.820567656
#> 89  089  0.22820828  0.209378054
#> 90  090  1.35177567  0.808692725
#> 91  091 -0.26031113 -0.848177138
#> 92  092 -0.80449599 -0.270387656
#> 93  093  1.20565591  0.016306548
#> 94  094 -0.91786101  0.298639347
#> 95  095  0.83334359  0.009773801
#> 96  096  1.30055384  0.214225224
#> 97  097 -0.54383485  0.875833909
#> 98  098 -0.84342658  0.288638963
#> 99  099  0.83079867 -0.590200792
#> 100 100  0.57429006  0.006835954