Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4597804
#> Y_2 0.4597804 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -2.2956912 -0.91829822
#> 2 -1.0537908 -0.05196657
#> 3 -0.2977659 -0.43204752
#> 4 -0.5402528  1.03647339
#> 5  0.6589067 -1.01307661

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1         Y_2
#> 1   001  0.2080008717 -0.81985160
#> 2   002 -0.7675723529  0.34025411
#> 3   003 -0.7545969309 -1.25621803
#> 4   004  1.2605643096  1.85929934
#> 5   005 -0.8855428239 -1.16640567
#> 6   006  1.3842465183 -0.89243457
#> 7   007 -0.0879624964  0.84074918
#> 8   008  1.4394919346  0.18052873
#> 9   009 -0.9751606035 -0.21110866
#> 10  010 -0.5835041500 -0.73475343
#> 11  011 -1.0055524313 -0.43646254
#> 12  012 -1.4215529716 -0.57053023
#> 13  013 -0.1282265349  0.78702400
#> 14  014 -2.5004517069 -1.36667028
#> 15  015  0.1902401485  0.07091694
#> 16  016 -0.4089079526 -0.25482023
#> 17  017  1.4475304460  0.72902008
#> 18  018  0.2143659737 -0.87172935
#> 19  019  0.1112326178 -0.06159145
#> 20  020 -0.2284156511 -1.71250442
#> 21  021 -0.4932061286 -0.73943327
#> 22  022  1.0197455528  1.66667679
#> 23  023  0.5714208907  1.29852952
#> 24  024 -0.5769753961 -0.65525731
#> 25  025 -0.2167201827  0.61255032
#> 26  026 -0.3755242964  0.71417269
#> 27  027  0.2157082973 -0.18431273
#> 28  028  1.2488081434  0.58144883
#> 29  029 -1.2467141177 -0.47263006
#> 30  030 -0.0472104213 -1.39861633
#> 31  031 -0.5686919395 -0.84707813
#> 32  032  1.0472704377 -0.17973304
#> 33  033 -0.5806095797  0.22813281
#> 34  034 -0.3887876511 -1.66553340
#> 35  035 -0.1487413978 -1.92195492
#> 36  036 -1.5317807861 -0.73954370
#> 37  037 -0.0641527021  0.09480412
#> 38  038 -0.9049502914  0.06230269
#> 39  039 -0.7619790215 -0.65228797
#> 40  040 -0.4598266345  0.18830122
#> 41  041  0.3287953522  0.24759172
#> 42  042 -1.2136287660 -1.43629576
#> 43  043 -1.4618261301 -0.02220579
#> 44  044  1.5061482787  1.18803564
#> 45  045  0.3864901522  0.71875896
#> 46  046 -1.3528458549 -1.81735538
#> 47  047 -0.4339345650 -0.05416060
#> 48  048 -0.8195075547 -0.57664882
#> 49  049 -1.3443887195 -0.11953272
#> 50  050 -0.9492314194 -0.27015679
#> 51  051 -0.9103327522 -0.57425605
#> 52  052 -0.3901835015 -0.27425013
#> 53  053  0.1920048513  0.64918282
#> 54  054 -0.9492617563 -0.59634922
#> 55  055 -0.4205857795  0.53429111
#> 56  056 -0.7393834860 -2.25365322
#> 57  057  0.7235816826  0.70504009
#> 58  058 -0.5789346091 -0.60882122
#> 59  059 -1.0362730181  0.50785227
#> 60  060  0.9851722296  0.79534665
#> 61  061  1.2993666199  1.62804447
#> 62  062  0.2545236821  0.46646155
#> 63  063 -0.6715376988 -0.59167949
#> 64  064 -0.5470148047 -0.64073098
#> 65  065  0.1553626157 -0.29556003
#> 66  066 -0.8144173709 -1.96695023
#> 67  067  2.1308924924  2.15650413
#> 68  068  1.0407382341 -1.05880925
#> 69  069 -0.0507798816  0.51132139
#> 70  070  1.2465079718 -0.08429868
#> 71  071  0.6182405101  0.12648684
#> 72  072  0.0795392067 -0.08716555
#> 73  073  0.4164981072 -0.61084272
#> 74  074 -0.5788841733  0.44658489
#> 75  075  1.5463626912 -0.06276968
#> 76  076  1.0397522808  1.11203291
#> 77  077 -0.4999280103 -1.58653556
#> 78  078 -0.4290643356 -1.20538526
#> 79  079 -0.4782031431 -0.07630845
#> 80  080  1.4617734296 -0.85944969
#> 81  081 -0.0770618651  1.56387552
#> 82  082  0.6797662876  1.08558643
#> 83  083 -0.5014956446 -0.40879729
#> 84  084 -1.4388631707 -0.65917113
#> 85  085 -0.8369484132 -0.84683703
#> 86  086  0.7071824982  0.38713872
#> 87  087 -0.7193000009 -1.72995160
#> 88  088  1.2188301793  1.64328588
#> 89  089 -0.1734431732 -0.54039700
#> 90  090 -0.3140199447 -0.41549731
#> 91  091 -0.1561193473  0.05390232
#> 92  092  1.2639219696  0.93838988
#> 93  093 -2.2178378254 -1.11307259
#> 94  094 -0.6697159672  0.53180958
#> 95  095  0.4210858469 -1.29897545
#> 96  096 -0.0009314433  0.94606011
#> 97  097 -0.1470674174 -0.72756243
#> 98  098 -0.8017887137  0.36665471
#> 99  099 -0.1974993413 -0.56895337
#> 100 100  0.8945736469  0.11676011

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.162449709  0.34051222
#> 2   002 -1.260821925 -1.16162913
#> 3   003  0.184376582  0.71457152
#> 4   004 -0.590292795  0.61566919
#> 5   005  1.104709105  0.48486338
#> 6   006  0.193802312 -2.17511665
#> 7   007 -1.033364049 -0.17030090
#> 8   008  0.115482458 -1.39110821
#> 9   009 -0.146168919 -0.33533796
#> 10  010 -0.949562965  0.29289644
#> 11  011  1.637855824 -0.40315252
#> 12  012  1.080419351 -0.27254326
#> 13  013 -0.584672021 -0.46786073
#> 14  014  1.135077260  0.94059849
#> 15  015 -0.873995426  0.15068382
#> 16  016 -0.399637373  1.05336713
#> 17  017 -0.574090920 -0.51604104
#> 18  018 -0.103808591 -0.31731466
#> 19  019  0.168056819 -0.80091615
#> 20  020 -0.555566913 -0.02424502
#> 21  021 -0.113043377 -0.14893393
#> 22  022 -0.263106599  0.12074584
#> 23  023  0.458337447 -0.07432898
#> 24  024  1.662087484  0.59700350
#> 25  025 -0.806792613 -0.57928348
#> 26  026  0.908973079  2.06616414
#> 27  027 -0.202670725  1.94369819
#> 28  028  0.579522910  0.56120092
#> 29  029 -0.684739326  0.36676279
#> 30  030 -0.786428564 -1.34352488
#> 31  031  0.617347225 -0.36787179
#> 32  032  0.338800026  1.32880049
#> 33  033  0.195996536  1.53019226
#> 34  034 -0.044907500 -0.88842104
#> 35  035  0.696798141  0.61893422
#> 36  036 -0.522631399 -0.59847783
#> 37  037  0.513884105 -0.08228652
#> 38  038  0.737954276 -0.20808819
#> 39  039  1.014495187  0.94116038
#> 40  040 -0.233904745 -1.48940476
#> 41  041 -0.176359877  1.14007634
#> 42  042 -0.201063707  0.52608428
#> 43  043  0.126461288 -0.60182264
#> 44  044 -0.019725609 -0.04487124
#> 45  045  0.049878334  0.05141806
#> 46  046 -0.982715244  0.13831568
#> 47  047  0.071290542 -0.65395619
#> 48  048 -0.401190058  0.20192539
#> 49  049 -0.802871111 -0.94096925
#> 50  050  1.405471334  0.97488960
#> 51  051 -1.397875547 -1.15604011
#> 52  052  1.048751779  1.64245164
#> 53  053  0.285917807 -0.36066316
#> 54  054 -1.052037077 -0.42364106
#> 55  055 -1.713911307 -0.78251477
#> 56  056 -1.102118921 -1.50485851
#> 57  057 -1.923242672 -0.24322712
#> 58  058  0.414043994  0.14722135
#> 59  059 -0.113654022  0.12364692
#> 60  060 -1.471121727 -0.56118923
#> 61  061 -0.288152278 -0.08907344
#> 62  062 -0.573596776 -1.01415580
#> 63  063  1.177373312  0.72642192
#> 64  064  0.909626931  0.08558465
#> 65  065 -0.955173144 -0.32094391
#> 66  066  0.157409221  0.29653250
#> 67  067  1.506713746  0.93373949
#> 68  068  0.235653906  0.27952431
#> 69  069  0.048048314 -0.80965511
#> 70  070 -0.119868706 -0.92971850
#> 71  071 -0.732038334 -1.35560526
#> 72  072  0.444332655  1.35929932
#> 73  073 -1.402545339  0.58642115
#> 74  074  0.001445081  0.24169688
#> 75  075 -1.149793440 -0.84613788
#> 76  076 -0.155366023 -0.05027219
#> 77  077 -0.297084106 -0.15838642
#> 78  078  0.426479330  0.38461272
#> 79  079 -0.227787986  0.84379015
#> 80  080  0.485357021 -0.63909780
#> 81  081 -0.020055904 -1.01250477
#> 82  082  0.832111439  0.12545809
#> 83  083 -0.442521733 -0.36589596
#> 84  084 -0.185812099  0.17404182
#> 85  085  1.082453157  0.57984649
#> 86  086 -0.580238658 -0.12344398
#> 87  087  0.479349542  0.72789685
#> 88  088 -0.306703627 -0.05979866
#> 89  089 -0.650018542  0.23661641
#> 90  090 -0.690820014 -0.21076555
#> 91  091 -0.595285878 -1.03658898
#> 92  092  0.930674685  2.80399456
#> 93  093 -2.295146288 -0.95719118
#> 94  094 -0.009758389  0.85031245
#> 95  095 -2.295219646 -1.43293452
#> 96  096  0.645517484 -0.13381433
#> 97  097 -1.598500382 -0.15182046
#> 98  098  0.161247790  0.86559217
#> 99  099 -1.047710924 -1.59238744
#> 100 100 -0.215655312 -1.22874626