Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4667649
#> Y_2 0.4667649 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.3214474 -1.2677651
#> 2  0.7212246  2.0308481
#> 3  1.0029831  0.5222471
#> 4 -0.5592093 -0.5343968
#> 5  1.5369708  0.4840865

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.05668164  1.136285269
#> 2   002  0.42294181 -0.580288163
#> 3   003 -0.29789184 -1.179336875
#> 4   004 -0.60985557 -0.730885245
#> 5   005  0.36360177  0.661395088
#> 6   006 -0.87813504 -0.695441049
#> 7   007 -0.78097324  0.004775255
#> 8   008 -0.37404290  1.184651860
#> 9   009 -0.55209062 -0.305127242
#> 10  010  1.51361898 -0.935994071
#> 11  011 -1.22327159  0.056193859
#> 12  012 -1.75594667  0.761223036
#> 13  013 -1.14457441  0.541463739
#> 14  014 -0.08329627  0.449546000
#> 15  015  0.69207668 -0.033974593
#> 16  016  0.94459090  1.843486814
#> 17  017 -0.21402646 -1.337477538
#> 18  018  0.03703184  0.214417956
#> 19  019 -0.01148497  0.960620624
#> 20  020  1.19013964 -1.167160315
#> 21  021  1.33827033 -0.280080999
#> 22  022  0.44419364 -1.291948126
#> 23  023 -1.72501302 -1.315276825
#> 24  024  1.97838520  2.370101391
#> 25  025 -0.99365859 -1.149740633
#> 26  026  1.79661280 -0.846235212
#> 27  027 -0.72095462  0.844832974
#> 28  028 -0.51199953  0.206894251
#> 29  029  1.09288530  0.813869153
#> 30  030  0.83158098  1.130082662
#> 31  031 -1.00972022 -1.843674697
#> 32  032  1.34761784  0.968960203
#> 33  033 -0.83705788 -1.924574764
#> 34  034 -0.39203195  1.105393476
#> 35  035  0.73008556 -0.636696931
#> 36  036  0.09278368 -0.252315108
#> 37  037 -1.18846825 -2.947355154
#> 38  038 -0.65031136 -0.517555825
#> 39  039  1.27679926 -0.115212743
#> 40  040 -0.05952295  1.513665132
#> 41  041 -1.19515878 -0.364335301
#> 42  042 -0.40411560 -0.801963494
#> 43  043  0.81283466 -0.552626533
#> 44  044 -0.08568195  0.457193990
#> 45  045 -2.48687800 -2.237801541
#> 46  046  0.08456948 -0.045072353
#> 47  047  1.44933944  1.000141850
#> 48  048  0.47646660 -0.980797036
#> 49  049 -0.02393742  0.076456511
#> 50  050  0.03767404  0.497526679
#> 51  051  0.73801209 -0.893774724
#> 52  052  0.55277261  0.784557237
#> 53  053  1.32286397  0.389399872
#> 54  054  1.23261843  1.099053843
#> 55  055 -0.65922747  0.499042443
#> 56  056  0.65733526 -0.195631615
#> 57  057 -0.78668905 -0.025544796
#> 58  058 -1.21854176 -1.189970568
#> 59  059  1.74611257  0.256734179
#> 60  060 -0.48119266 -0.610269985
#> 61  061  0.49199515 -0.104589142
#> 62  062 -1.19379518  0.643269607
#> 63  063  0.43058852  1.048329965
#> 64  064  0.40180386  0.992205909
#> 65  065 -1.85491820 -0.806384941
#> 66  066 -0.12715240 -0.986590359
#> 67  067  1.13079325  0.405252195
#> 68  068 -1.95168061 -1.136787619
#> 69  069  0.39549271 -0.018686304
#> 70  070 -1.24023408  1.562783211
#> 71  071 -0.51494327  1.181207910
#> 72  072 -0.13006603 -0.558431227
#> 73  073  0.31986704 -0.067357485
#> 74  074  0.65114985 -0.262314615
#> 75  075  1.06466157 -0.135786594
#> 76  076  1.74507343  0.525858208
#> 77  077  0.76037524 -0.080123795
#> 78  078  0.46598519 -0.025587757
#> 79  079 -0.72964187 -0.624152949
#> 80  080 -0.27659955 -1.073817755
#> 81  081  0.30321844 -0.712140401
#> 82  082  1.42556769  1.090409752
#> 83  083  1.07736932 -0.004939965
#> 84  084  0.14584609  0.500396757
#> 85  085  1.28287638  0.593294406
#> 86  086  0.50937401  0.012024699
#> 87  087 -1.35176185 -0.117731445
#> 88  088 -0.54358828 -0.564999567
#> 89  089 -1.49803147  0.031754175
#> 90  090  0.15464414 -0.619446519
#> 91  091  0.03541825 -1.419761100
#> 92  092  2.28070130  0.319354035
#> 93  093  1.42823163  1.684447861
#> 94  094 -0.57923147 -0.008332679
#> 95  095 -0.02888131 -1.249063134
#> 96  096  0.95887567  1.015296102
#> 97  097  0.39165919  1.689690711
#> 98  098  1.09976264  1.335728237
#> 99  099  0.10208205 -0.332594651
#> 100 100 -0.15642416  0.951903244

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.503877592 -1.86134852
#> 2   002  0.347866778  2.08533900
#> 3   003 -0.321234174  0.09556793
#> 4   004 -0.559991421 -0.13941738
#> 5   005  2.903240668  1.97496010
#> 6   006  0.624278009  0.49212822
#> 7   007 -0.650994264 -0.57497991
#> 8   008  0.587402036  0.73847974
#> 9   009  0.699202521  1.56518770
#> 10  010  0.620131002  1.00797784
#> 11  011  0.805490438  0.65637125
#> 12  012  0.691872347  1.46282624
#> 13  013 -0.067596609 -0.95683385
#> 14  014 -0.619125689  0.88205397
#> 15  015 -0.710994607 -0.23852582
#> 16  016 -0.953511831 -0.03032196
#> 17  017 -1.120496231  0.13432437
#> 18  018 -2.431651704 -0.08426312
#> 19  019 -0.586904077 -0.02619858
#> 20  020  0.696913391  1.01236367
#> 21  021 -0.006612841  0.85724965
#> 22  022 -0.362460589  0.74906060
#> 23  023  2.355190207  1.30925828
#> 24  024 -0.810371560 -0.69433705
#> 25  025 -0.912566052 -1.05576164
#> 26  026 -1.311784261  0.96170404
#> 27  027  1.211356456  0.55398190
#> 28  028 -1.912110174 -2.21890637
#> 29  029 -0.492875177 -0.31237062
#> 30  030 -1.486933997  0.56548616
#> 31  031 -1.101782509  0.15894643
#> 32  032  0.795348505  0.22142225
#> 33  033  0.727866863  0.63183318
#> 34  034  0.099063441 -1.32809700
#> 35  035 -0.137756248  1.24879814
#> 36  036  0.832209000 -0.25490176
#> 37  037  0.665333544 -1.22507123
#> 38  038  0.975455398 -1.41964626
#> 39  039 -1.135280069 -0.38833771
#> 40  040 -0.513188316 -0.46345282
#> 41  041 -0.119991814  0.38549847
#> 42  042  0.548799150 -0.09087907
#> 43  043  0.101200112  1.43075183
#> 44  044  0.298027563  0.49543005
#> 45  045 -0.566986213  1.16338338
#> 46  046 -1.094784092  0.56857106
#> 47  047  0.790474127  1.66133575
#> 48  048  0.283125708 -0.80745202
#> 49  049  0.353401784  0.50693237
#> 50  050  1.894848642 -0.19018631
#> 51  051  0.879070612  0.41833088
#> 52  052  0.116029863 -0.70002743
#> 53  053 -1.000031656 -0.23340319
#> 54  054 -0.943391424 -0.78314195
#> 55  055  0.297528020  1.10953740
#> 56  056  0.032147905 -2.04088887
#> 57  057  0.130795168  0.54795998
#> 58  058  0.824868826 -1.42362983
#> 59  059 -0.058861362  0.45884713
#> 60  060  2.301939247  0.75373058
#> 61  061 -0.712756813  0.46854824
#> 62  062  0.543535173  0.74207285
#> 63  063  0.855468677  0.16108203
#> 64  064  1.063924229 -0.15689286
#> 65  065  1.608461894 -0.42572475
#> 66  066 -0.216115008 -1.14516492
#> 67  067 -0.818448886 -0.40101460
#> 68  068 -0.300070989  0.06303850
#> 69  069  0.192801139  0.30757823
#> 70  070  0.138352123 -1.92607376
#> 71  071 -0.488586625  0.41970485
#> 72  072  0.273479314 -0.01687191
#> 73  073 -2.086718539 -1.84850410
#> 74  074  0.252887327  0.48446757
#> 75  075  1.279336469  1.29199294
#> 76  076 -0.444157346  0.70717711
#> 77  077 -0.545994516  0.09923551
#> 78  078  0.759326515  0.63613495
#> 79  079  0.223086703  0.58790209
#> 80  080 -0.002514944  0.35297380
#> 81  081 -0.327492220 -1.21403581
#> 82  082  1.656017692  1.14612598
#> 83  083 -0.636160685  0.69524675
#> 84  084  1.353678952 -0.23300745
#> 85  085  1.604686235 -0.06618422
#> 86  086  0.147621057 -0.46154184
#> 87  087  0.602789650  0.71570279
#> 88  088  1.746127708  0.46973659
#> 89  089  1.491459049  0.38885339
#> 90  090  1.059556238  0.02296441
#> 91  091  2.243865012  0.61016409
#> 92  092  1.762734072  1.51200957
#> 93  093  1.948001117  0.97782536
#> 94  094  0.589465651  1.09498495
#> 95  095 -1.144297045 -0.25509069
#> 96  096 -0.226271748  0.48181579
#> 97  097 -0.490660666 -0.79345921
#> 98  098 -0.172637561 -0.45191985
#> 99  099  0.139835987 -1.49420232
#> 100 100  1.259662050  2.80640915