Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5218478
#> Y_2 0.5218478 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.7047193  1.0593342
#> 2  1.3305957  0.2242521
#> 3  0.6140038  0.2554350
#> 4 -0.8929172  0.4424106
#> 5 -0.3288515 -0.7802902

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.301683248  0.133491396
#> 2   002  1.558976248  2.414152040
#> 3   003  0.033432908 -0.470283563
#> 4   004  1.079154465  1.495227276
#> 5   005  1.254065844  1.679124926
#> 6   006 -0.640182820  0.629095228
#> 7   007  1.961241512  1.884297769
#> 8   008  0.738820698  0.420709243
#> 9   009 -0.132888226  1.345110326
#> 10  010  0.774221503  2.436977488
#> 11  011  0.366268901  0.479723228
#> 12  012  0.276006775  0.759377614
#> 13  013  0.621683565  1.090469994
#> 14  014  0.722498996 -0.676451166
#> 15  015 -1.130619584 -1.639138690
#> 16  016 -0.557351030  0.537306303
#> 17  017 -0.223224665  0.135009460
#> 18  018  0.633392501  0.893888173
#> 19  019  0.387065415 -0.788859792
#> 20  020 -1.281704409 -1.099767988
#> 21  021 -0.404152040 -0.004371990
#> 22  022 -0.704619855 -0.964258760
#> 23  023 -0.842552820 -0.451282684
#> 24  024  0.231978864  0.815788135
#> 25  025 -0.098750637 -1.401821954
#> 26  026  1.093311903  1.526716827
#> 27  027 -1.419346816 -0.755139818
#> 28  028 -0.212099392  0.461308208
#> 29  029  0.229337301 -0.631943329
#> 30  030  0.727542235  0.331863412
#> 31  031 -1.096454481 -1.144168638
#> 32  032 -0.291079464  1.575441005
#> 33  033 -1.338261877 -0.662445584
#> 34  034  1.293520034 -0.136778866
#> 35  035 -2.138138703 -1.984586272
#> 36  036 -1.220834460 -1.424875699
#> 37  037 -2.125803140 -0.155065543
#> 38  038  0.554115326 -0.684061694
#> 39  039  1.070124475  2.327497720
#> 40  040  0.197538321 -0.549784480
#> 41  041  0.076780597  0.211578214
#> 42  042  0.970781523  1.844447533
#> 43  043  0.782967296  0.180617791
#> 44  044 -0.552774841 -0.709766043
#> 45  045  0.586146621  0.626924889
#> 46  046  0.207039370 -0.617752140
#> 47  047 -0.743478509 -1.127827959
#> 48  048  1.412920478  1.126977516
#> 49  049 -0.546638724 -0.927121443
#> 50  050 -0.729377414 -0.700440677
#> 51  051 -1.473821477  0.105638634
#> 52  052  1.237947217 -0.326499488
#> 53  053 -0.147478776 -0.776662677
#> 54  054 -0.789077774 -0.591673031
#> 55  055  0.386330766  1.687455990
#> 56  056  0.726897615  0.007697227
#> 57  057 -0.148300752  0.262224449
#> 58  058  1.369350995  1.318892202
#> 59  059  0.169971484  0.374875517
#> 60  060  0.781113217  1.964278722
#> 61  061  1.589461931  2.107615103
#> 62  062  0.287040113 -1.131180532
#> 63  063  0.023668688 -2.039062890
#> 64  064  0.059835950 -0.644530234
#> 65  065  1.046669128  1.531114619
#> 66  066 -0.116307188 -1.130919746
#> 67  067  0.837725853  1.238182885
#> 68  068  0.774612860  1.003161589
#> 69  069 -0.618105500  0.224722577
#> 70  070  0.748523951  1.729471014
#> 71  071 -0.970369979 -0.629463624
#> 72  072  1.505132212  0.127706220
#> 73  073 -0.246648544  0.745795040
#> 74  074  0.008178373 -0.294494837
#> 75  075  2.552497556 -1.176570888
#> 76  076 -0.190300450 -0.131110545
#> 77  077 -0.725624929  0.796029701
#> 78  078 -1.183886948 -0.192275756
#> 79  079  0.272823730  1.096519426
#> 80  080 -0.330051447  0.213162219
#> 81  081 -0.190382757  0.518947189
#> 82  082 -0.415708043 -0.508025253
#> 83  083 -0.603207164 -0.405799454
#> 84  084  0.228195245 -1.547362110
#> 85  085 -1.288568607 -1.863861566
#> 86  086 -0.287004981 -0.036655908
#> 87  087 -1.400999963 -0.794916032
#> 88  088 -1.008516793 -0.365932761
#> 89  089 -0.308437613 -0.650801498
#> 90  090 -1.860923345  0.254723442
#> 91  091  0.306840443  0.066737516
#> 92  092 -0.152412016 -0.741033234
#> 93  093 -0.079950235  0.309622890
#> 94  094  0.379063480 -0.792227498
#> 95  095 -0.497699084 -0.715312723
#> 96  096  0.458077765  1.115123860
#> 97  097 -1.033969712 -0.169908212
#> 98  098 -0.313313247  0.889334295
#> 99  099 -0.617531312  0.926836022
#> 100 100 -1.755690013  0.209584289

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.223916422  0.10950414
#> 2   002 -0.511192452  0.04174270
#> 3   003  1.301064790  1.73061590
#> 4   004 -0.049272692  0.17710814
#> 5   005 -0.002481707  0.19815577
#> 6   006  1.222812522  0.38832187
#> 7   007  0.850999528 -1.50823372
#> 8   008  1.090186589 -0.39367867
#> 9   009 -1.118951053  0.10286917
#> 10  010  1.630134714  0.64457704
#> 11  011  0.541358627  2.36606780
#> 12  012 -0.762652544 -1.90375904
#> 13  013  0.542978599  0.69848234
#> 14  014  1.517494181  1.14717800
#> 15  015  0.764393615  1.41033471
#> 16  016  1.568858692  0.37855571
#> 17  017 -0.047559972  0.13687077
#> 18  018 -0.363726724  1.06486947
#> 19  019  0.727108053  0.82381960
#> 20  020 -0.980977795 -1.09320051
#> 21  021  0.179774404 -0.01985786
#> 22  022 -1.522633054 -1.70380363
#> 23  023 -2.162487074 -0.24572005
#> 24  024 -0.154060539 -1.90680811
#> 25  025  0.532181657  0.03346211
#> 26  026  2.510543137  1.28679252
#> 27  027  0.021183155  0.85034515
#> 28  028 -0.914705882 -0.69163991
#> 29  029  1.812699547  1.36489942
#> 30  030 -1.007107629 -0.11698829
#> 31  031  0.490826518  0.18245537
#> 32  032 -1.387370549 -0.51825842
#> 33  033  0.793953052  0.09993112
#> 34  034 -0.606972866  0.39601642
#> 35  035 -0.616689311 -1.02191760
#> 36  036  1.572746605  1.51776927
#> 37  037  1.025443720 -0.59777563
#> 38  038  0.559766817  1.05787174
#> 39  039  1.651441370  1.42911012
#> 40  040 -0.647451550  0.66579059
#> 41  041  0.671323781 -0.32111391
#> 42  042  0.138904001 -0.76127988
#> 43  043  0.205950255  0.01298068
#> 44  044  0.129237789 -0.29115631
#> 45  045 -1.096147312 -2.89995908
#> 46  046  0.419387281  0.99290487
#> 47  047  0.934751752  0.00994930
#> 48  048 -0.447987058  0.25135253
#> 49  049  0.143624845 -0.14390059
#> 50  050 -0.807436941 -0.34985945
#> 51  051 -1.076156932  0.55768069
#> 52  052 -1.885737985 -1.68720059
#> 53  053 -0.742740516  0.59740001
#> 54  054 -1.123374554  0.31467693
#> 55  055 -0.540826032  0.18968321
#> 56  056 -0.172719871 -0.71629609
#> 57  057  0.185552315  2.39961923
#> 58  058 -1.957958118 -2.66025900
#> 59  059  1.419183190 -0.07703781
#> 60  060  0.629103400  0.82271980
#> 61  061 -0.589515835  1.58983042
#> 62  062 -0.930626119  0.21062884
#> 63  063  0.400687267  0.30245859
#> 64  064  1.286501306  0.17817256
#> 65  065  0.331870995  0.50980077
#> 66  066 -1.651528426 -0.26908407
#> 67  067 -1.487058744 -2.35578453
#> 68  068 -0.863427598  0.15687478
#> 69  069 -0.399563166 -2.21652077
#> 70  070  0.280258881 -0.26105336
#> 71  071 -0.281229799  0.43591136
#> 72  072  0.602317052  0.56733172
#> 73  073 -1.293466511 -0.68441627
#> 74  074 -0.008750763 -1.00514650
#> 75  075 -0.347870581 -0.61746340
#> 76  076 -0.704159964 -1.51475581
#> 77  077  0.473433319  0.72476883
#> 78  078 -0.075718607 -0.67357872
#> 79  079 -1.202221809 -1.64106504
#> 80  080 -1.725862968 -1.22991446
#> 81  081 -1.109852920 -0.24859747
#> 82  082  0.522651059 -0.83325718
#> 83  083 -1.110919835 -0.74377210
#> 84  084  1.611972445  1.00898985
#> 85  085  0.243904273 -0.24640870
#> 86  086  1.121730766  1.21970031
#> 87  087 -0.156532620  0.34961305
#> 88  088  0.544071595  1.85556613
#> 89  089  0.272256999  0.19324284
#> 90  090 -0.690808075 -1.64970166
#> 91  091  1.125646850 -0.47856821
#> 92  092 -1.374795850 -1.84681218
#> 93  093  0.296990876 -0.59799004
#> 94  094 -0.821815672 -1.22258566
#> 95  095  0.755455140 -0.26224123
#> 96  096  0.998650904 -1.20292633
#> 97  097  1.446143413  0.99791740
#> 98  098 -1.759515512 -0.90280591
#> 99  099  0.074685680  0.24586968
#> 100 100  1.091103082  0.74685370