Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4729002
#> Y_2 0.4729002 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3810532 -0.8413095
#> 2 -0.2558038 -0.3661890
#> 3  1.9294105 -0.4305334
#> 4 -1.1073650  0.9163194
#> 5 -0.5072748 -0.3308702

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.04207504 -0.02631976
#> 2   002  2.56385484  1.90098970
#> 3   003  0.18758287 -0.45322033
#> 4   004  1.10544947  1.82036335
#> 5   005 -0.88896253  0.91901172
#> 6   006 -1.15471608 -0.03990951
#> 7   007 -0.52646331 -0.05789385
#> 8   008  1.89710276  0.19562682
#> 9   009 -0.16071525  0.95868771
#> 10  010  0.59209864 -0.60661756
#> 11  011 -0.44319174  0.19314066
#> 12  012  2.01771235  0.61703612
#> 13  013  0.59768329 -0.15952050
#> 14  014  0.68767416  1.09594439
#> 15  015  0.28501120 -0.38968368
#> 16  016  0.77984622 -0.43681449
#> 17  017 -1.43642340 -0.76952743
#> 18  018 -1.33436218 -0.93987575
#> 19  019  1.10140950  0.01368051
#> 20  020  0.34588707  0.80997491
#> 21  021 -2.90380987 -0.88680430
#> 22  022 -0.46325850 -0.23072866
#> 23  023 -0.38272835 -0.71427078
#> 24  024 -0.57191275 -0.11133549
#> 25  025  0.17042837  0.82354653
#> 26  026  0.55032920  0.33160773
#> 27  027 -0.10060622  0.42653343
#> 28  028 -0.42897292  0.42037270
#> 29  029 -0.12600511 -0.36831843
#> 30  030  1.19307178  1.34816382
#> 31  031  1.58438952  0.71690525
#> 32  032 -0.08452980 -0.61050384
#> 33  033 -0.75603878 -1.53607523
#> 34  034 -0.66566488 -0.16987967
#> 35  035  0.02697040  1.21057691
#> 36  036 -2.22795052 -1.76021730
#> 37  037 -0.02600163 -0.92954802
#> 38  038 -0.59332744 -1.03169351
#> 39  039  0.29679393  0.15883351
#> 40  040 -0.34289906  1.07329877
#> 41  041  0.27572422  0.46766327
#> 42  042 -0.84063973  0.41297228
#> 43  043 -0.55516845 -1.13937641
#> 44  044  0.79681257  0.75699351
#> 45  045  1.69402594  2.08033331
#> 46  046  1.03419227  1.87135731
#> 47  047  0.82755371  0.45948845
#> 48  048 -0.51422543 -0.29383914
#> 49  049  0.56895520  0.73146258
#> 50  050 -0.54089430 -0.49042574
#> 51  051 -0.42668788 -0.33024209
#> 52  052 -2.15965813 -0.83790940
#> 53  053  0.21535197 -0.36211436
#> 54  054 -0.49109368  0.12368548
#> 55  055 -0.65843612 -0.12277944
#> 56  056 -0.47152535  0.67019409
#> 57  057  0.25828142 -0.97485160
#> 58  058  1.28263115 -0.52985699
#> 59  059 -0.45327158  0.02016606
#> 60  060 -0.08018231 -0.05528778
#> 61  061  1.04792540 -1.71847133
#> 62  062 -0.61012430 -0.44745293
#> 63  063  0.42573613  1.75597261
#> 64  064  1.54446045  0.16781901
#> 65  065 -1.65714516  0.09397169
#> 66  066  1.30269807  0.52801150
#> 67  067  1.86024803  1.14131570
#> 68  068  0.30167372 -0.28198578
#> 69  069  1.35193256  0.89647527
#> 70  070  1.42756272  1.09860529
#> 71  071  0.86769812  0.93207581
#> 72  072 -0.76994525 -0.15075997
#> 73  073 -0.29887825 -1.54851328
#> 74  074 -1.57319418 -1.09468750
#> 75  075 -2.47523036 -1.44984847
#> 76  076 -0.33509746 -2.20210275
#> 77  077 -0.89341687  0.86600306
#> 78  078 -0.81311854  1.10652700
#> 79  079  1.97301761  0.12952109
#> 80  080 -2.38387590 -2.18993383
#> 81  081 -0.63401866 -0.31671013
#> 82  082  1.91324203  1.66740099
#> 83  083  1.87333718  0.82685700
#> 84  084  0.22102511  0.68893426
#> 85  085  0.50595808  0.38230958
#> 86  086 -0.09526384  0.87655974
#> 87  087 -0.38786420  0.73566117
#> 88  088 -1.76363271 -0.93376404
#> 89  089 -0.36324547  0.42465801
#> 90  090  1.03504570  1.03579680
#> 91  091  2.14743486  2.47450314
#> 92  092  0.35633067  0.60075550
#> 93  093 -0.40464567  0.19918732
#> 94  094  0.29925028  1.16657436
#> 95  095  0.47083813 -0.62128863
#> 96  096 -0.93926751 -0.71420864
#> 97  097  0.19682654  0.18199389
#> 98  098  0.97767577 -0.04733852
#> 99  099  0.87667014  0.49188693
#> 100 100  0.74304390  2.48293277

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  1.27279833 -0.690759422
#> 2   002 -0.78123738 -1.294944622
#> 3   003 -1.73223176 -0.525263305
#> 4   004 -0.05267080  1.042670676
#> 5   005  1.35103292  0.294967656
#> 6   006 -1.18880071 -0.956419774
#> 7   007 -0.12673590 -0.181427547
#> 8   008  1.25617059  1.362271859
#> 9   009 -0.60167155 -1.610087275
#> 10  010  1.05091825 -0.438526881
#> 11  011  0.54327490 -0.959822121
#> 12  012 -0.35103718  0.481060477
#> 13  013 -0.46337677  0.443800445
#> 14  014 -0.67961639  0.008538669
#> 15  015 -0.01125544  1.423827065
#> 16  016 -0.98649343 -3.095440095
#> 17  017 -0.35050734  0.212483815
#> 18  018 -0.06535614 -0.134979611
#> 19  019  0.74698606 -0.671630410
#> 20  020  0.68227607 -0.222138654
#> 21  021  0.49871608  2.447018561
#> 22  022  1.38208030  1.223718420
#> 23  023  0.19754673  1.889196239
#> 24  024  3.31062126  1.606832365
#> 25  025  0.67249085  1.103422529
#> 26  026  0.89874299 -0.167883233
#> 27  027 -0.44991961  1.197541876
#> 28  028  1.56732893  2.184982127
#> 29  029  1.42194119  0.685809816
#> 30  030  0.47938116  0.761402874
#> 31  031 -0.25470747  0.650203917
#> 32  032  0.89530125  0.231967618
#> 33  033 -0.34769700 -0.615646976
#> 34  034  0.87762571  0.594421234
#> 35  035  1.11242239  1.760824122
#> 36  036  0.67060681  0.184860419
#> 37  037 -0.52780200  0.067162203
#> 38  038  0.54130113 -0.497802738
#> 39  039 -0.42880067 -0.360072326
#> 40  040 -0.77940705  0.072298429
#> 41  041  0.14808909  0.053374525
#> 42  042  1.42368084  0.447682995
#> 43  043 -0.71213359 -0.052934074
#> 44  044 -0.73695500 -1.662437457
#> 45  045  0.25639640  0.201546184
#> 46  046  0.53706630  1.073329891
#> 47  047 -1.05651656 -1.649750057
#> 48  048 -1.06937361 -0.117226000
#> 49  049  0.21488509  1.257024523
#> 50  050 -0.65448210 -0.352999113
#> 51  051 -0.42159032 -0.323048771
#> 52  052 -1.01933074  0.517659973
#> 53  053  0.42481407  1.499106785
#> 54  054  0.67980581  1.735124820
#> 55  055 -1.83322801 -1.947125334
#> 56  056 -0.12550097  0.713301080
#> 57  057  1.82582096  0.611274996
#> 58  058 -0.43063716  0.245134729
#> 59  059  0.55086059 -1.038385851
#> 60  060 -0.48714643 -0.509069322
#> 61  061 -1.34334977 -0.510233335
#> 62  062  0.46608022  1.014582365
#> 63  063 -0.41152712 -1.015451092
#> 64  064 -0.52208939 -0.898788420
#> 65  065  1.04510257  2.043964593
#> 66  066  0.91906285  1.573710782
#> 67  067  0.78743125 -0.080574059
#> 68  068 -0.65585617 -0.297882966
#> 69  069  0.53079540  0.158339396
#> 70  070 -0.59564963 -0.543975632
#> 71  071  1.29288662  0.214700366
#> 72  072  0.02891017 -0.619915540
#> 73  073  0.11346681 -0.301991104
#> 74  074 -1.63668190 -0.396200514
#> 75  075 -0.69624251 -0.473062221
#> 76  076 -0.46663287  0.106777300
#> 77  077 -0.49553097  0.979753941
#> 78  078 -0.55307920  0.442636158
#> 79  079  0.84653504 -0.260816268
#> 80  080  0.16452472  0.223934376
#> 81  081 -0.36331607 -0.833470007
#> 82  082  0.88742164 -0.231330175
#> 83  083 -2.25142099 -0.924181417
#> 84  084  1.52138863  0.722779295
#> 85  085 -1.40321315  0.662507645
#> 86  086  1.17306858  0.181344819
#> 87  087 -0.75082282 -2.058872171
#> 88  088 -2.14496584 -1.084034472
#> 89  089  0.67551971  0.267936085
#> 90  090 -0.25654167 -0.552714721
#> 91  091 -1.06681087 -0.182628870
#> 92  092  0.63673529 -0.606676070
#> 93  093  0.12943921  1.162401628
#> 94  094 -0.55801944  0.059263853
#> 95  095 -0.56696798 -0.315719319
#> 96  096  1.44915980 -0.619265431
#> 97  097 -0.25602614  1.552640020
#> 98  098 -0.55639786  0.069154487
#> 99  099 -0.50593288  0.090008645
#> 100 100  0.51778657  0.985016944