Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4537232
#> Y_2 0.4537232 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2070785 -1.4473525
#> 2 -0.6147621 -0.7353766
#> 3 -0.6880267 -1.0135204
#> 4  0.3651831  1.6514022
#> 5 -0.7916992 -0.4215159

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.45909657 -0.357423279
#> 2   002 -0.86127933  0.276517654
#> 3   003  0.29477853  0.176911754
#> 4   004 -2.08939911 -0.922735654
#> 5   005  0.61876805  1.297133623
#> 6   006 -0.26494394 -0.491335845
#> 7   007 -0.55241741 -0.392953372
#> 8   008  0.90338590  0.872547885
#> 9   009  0.31714017 -0.057951707
#> 10  010  0.34013979 -1.663602383
#> 11  011 -0.29220414 -0.641948122
#> 12  012 -0.43020057 -1.978264426
#> 13  013 -1.33518666 -3.247256954
#> 14  014  0.81434516 -0.800702500
#> 15  015  0.84616728  1.309229141
#> 16  016 -1.37308464 -1.620883704
#> 17  017  0.18918790  0.246069108
#> 18  018  0.49527637 -1.598450109
#> 19  019  0.09484190 -0.145874405
#> 20  020  0.48940048  0.297001756
#> 21  021 -1.57758037  1.091065112
#> 22  022 -0.76818294 -0.727557203
#> 23  023  1.55360227 -0.574259750
#> 24  024  1.63072621 -0.021284541
#> 25  025 -1.36197328  0.344005391
#> 26  026  0.92635864  1.178389257
#> 27  027 -0.96603364 -0.703123750
#> 28  028 -0.49177190 -0.290282977
#> 29  029  0.14785138  0.278179707
#> 30  030 -0.77883345 -1.058998737
#> 31  031  1.11232794  0.673698008
#> 32  032 -0.22961347 -0.335688576
#> 33  033  1.74056095  1.557243620
#> 34  034 -0.49935695 -0.610427310
#> 35  035  0.04704133 -0.134017425
#> 36  036  0.24711755  1.022885109
#> 37  037 -1.37025669 -1.761791031
#> 38  038  0.10569862  0.221842943
#> 39  039 -0.97404433 -0.887328609
#> 40  040  1.03782867 -0.462316612
#> 41  041 -0.35234823  0.862821203
#> 42  042  1.42924435 -0.053996405
#> 43  043 -1.30049330 -1.650573776
#> 44  044 -0.38656337  0.015633221
#> 45  045 -0.23076729 -1.024082887
#> 46  046 -0.54394225 -0.659283356
#> 47  047 -0.14414554 -1.153987677
#> 48  048  1.24516874  0.904726812
#> 49  049  0.59200839  0.328434834
#> 50  050 -0.10830686 -0.329999925
#> 51  051  0.26115361  1.212774667
#> 52  052  1.87521143  0.078518147
#> 53  053 -1.14527902  0.593036662
#> 54  054 -1.23802787  1.169261100
#> 55  055 -0.84548404  0.996576131
#> 56  056  0.71309156  1.177964286
#> 57  057 -0.21859251 -0.160394416
#> 58  058 -0.79949241  0.130357967
#> 59  059  0.45194392 -1.707982597
#> 60  060  0.46129268  0.791936465
#> 61  061 -1.14956438  1.045705756
#> 62  062 -1.34832836 -0.776189905
#> 63  063 -1.69439658 -2.029625866
#> 64  064  0.86185939  1.201798282
#> 65  065 -1.15156517 -0.276800556
#> 66  066 -1.58101357 -0.515597929
#> 67  067  1.20731619  1.201511717
#> 68  068 -0.30292951 -0.803452495
#> 69  069 -0.47479966  0.280019736
#> 70  070 -0.38134045  1.966792235
#> 71  071  0.04979545  0.710652813
#> 72  072 -1.51215169  0.474544073
#> 73  073  0.11099473 -0.648299988
#> 74  074 -0.46669945  0.347972763
#> 75  075 -0.60166191  0.983994559
#> 76  076 -1.13559182 -1.522456372
#> 77  077 -0.18111097  0.067688164
#> 78  078 -0.05826019 -0.646771404
#> 79  079 -1.45469960 -0.877380552
#> 80  080 -1.33116905 -0.135895725
#> 81  081  1.19950693  1.241924879
#> 82  082 -0.94800681 -1.976089313
#> 83  083 -1.66171397 -1.453362516
#> 84  084  1.19306001  0.792414351
#> 85  085 -0.28085147 -1.004297822
#> 86  086  2.19576019  0.760332197
#> 87  087 -0.94329105 -0.420738222
#> 88  088  0.27759622  2.040030860
#> 89  089 -0.34425885 -1.019108885
#> 90  090  0.19926754 -0.132411246
#> 91  091  1.94805167  0.008587151
#> 92  092  0.30653341 -0.107825335
#> 93  093 -0.57104866 -0.791782296
#> 94  094  0.76356214  0.545574248
#> 95  095  0.17680439 -0.423384532
#> 96  096  0.26917745 -0.597733350
#> 97  097  0.05608493  0.266871097
#> 98  098  1.28589184 -0.011823271
#> 99  099  0.28398553 -1.003713799
#> 100 100 -0.87065834 -2.547858718

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.53458297 -0.56957807
#> 2   002 -0.24457847 -0.51245338
#> 3   003 -1.20131158  0.64195547
#> 4   004  1.08843589  0.62591370
#> 5   005 -2.10078681 -1.01598054
#> 6   006 -1.00425912 -0.56400493
#> 7   007 -0.12627189 -0.71893293
#> 8   008 -0.83021173 -0.13575752
#> 9   009 -0.84529551 -2.03256963
#> 10  010 -0.49454574  0.86031806
#> 11  011  0.72905345  1.53472115
#> 12  012 -0.81575123 -0.53440041
#> 13  013 -1.69276854 -0.93276261
#> 14  014  0.84301179  0.49310493
#> 15  015 -0.66743228 -0.91667947
#> 16  016  0.30537151  1.42980354
#> 17  017  0.93292824  0.51457492
#> 18  018 -0.95721543  0.34434246
#> 19  019 -1.05477221 -0.42907403
#> 20  020 -1.35364691  0.03181686
#> 21  021  2.10430381  0.36518941
#> 22  022 -1.13339407 -1.42151472
#> 23  023  1.34576041 -1.30590850
#> 24  024  0.39818918  1.06134075
#> 25  025 -0.01284414 -0.01437903
#> 26  026 -1.31412738 -0.92458574
#> 27  027 -1.56771092 -0.70553223
#> 28  028  0.51874135  0.22341607
#> 29  029 -0.44419543  0.56395041
#> 30  030 -1.85774023 -0.65355833
#> 31  031  0.31044545  0.15735970
#> 32  032  0.12691364 -0.33018459
#> 33  033 -1.80872098  0.45262798
#> 34  034 -0.23398874 -1.41897345
#> 35  035  0.07538111  0.39996497
#> 36  036  1.03224635  0.66384089
#> 37  037 -0.39427134 -0.80583018
#> 38  038 -1.44605874 -0.32760223
#> 39  039  0.95506524  1.23352470
#> 40  040 -0.90901203  0.63892786
#> 41  041  0.10925636  0.06401217
#> 42  042 -1.02487505 -1.80965596
#> 43  043 -0.02858310  0.32059111
#> 44  044  0.02208282 -0.27103619
#> 45  045 -0.22209808  0.22023053
#> 46  046  0.84249809  1.51270108
#> 47  047 -1.10292749 -0.81794889
#> 48  048 -0.25398253  0.61597855
#> 49  049 -0.93179103 -1.30685568
#> 50  050  0.42932110  1.49966274
#> 51  051  0.44114997  1.54270724
#> 52  052  0.94783732 -0.29995007
#> 53  053  1.60969161  1.18660762
#> 54  054  1.62264381 -0.54844028
#> 55  055 -0.79445177 -0.53081330
#> 56  056 -1.27804277 -1.81739974
#> 57  057 -0.70861920  0.72277333
#> 58  058  0.72203406  0.29539427
#> 59  059 -1.80496689 -0.12547166
#> 60  060  0.97842416  0.35112496
#> 61  061  1.92749999  2.73029175
#> 62  062  0.43096865  0.82451406
#> 63  063 -0.44300693 -0.55395944
#> 64  064  1.37371184  0.94581321
#> 65  065 -1.18026495  0.75945032
#> 66  066 -0.13297356 -1.51094118
#> 67  067  0.39986231  0.83418775
#> 68  068 -0.59758362  0.17791183
#> 69  069  0.21698437 -1.01432206
#> 70  070 -0.59436061 -0.77024397
#> 71  071  0.33419685  1.15210484
#> 72  072  0.04226289 -0.62887605
#> 73  073  0.47899525  0.58724085
#> 74  074 -0.67366518  0.55764334
#> 75  075 -0.27051303 -1.53732756
#> 76  076 -0.40500270  0.53369664
#> 77  077  0.48411774 -1.51579902
#> 78  078 -1.56120546 -0.34137769
#> 79  079  0.21427682 -0.62287961
#> 80  080  0.87827284  0.82691498
#> 81  081 -0.72743546 -0.09794386
#> 82  082  0.52268379  0.23517370
#> 83  083 -0.33303128 -0.10417257
#> 84  084 -1.35421543 -1.89294379
#> 85  085  0.57163192 -0.06928789
#> 86  086 -0.68885581  1.06589449
#> 87  087 -1.42127017 -1.47854193
#> 88  088  0.89848651 -0.49095072
#> 89  089 -0.49636260 -0.39557783
#> 90  090  0.07761282 -0.46474985
#> 91  091 -1.44599573 -0.51484655
#> 92  092  0.96669803  0.33926249
#> 93  093 -0.21838808 -0.15922637
#> 94  094  0.24196456  0.48786181
#> 95  095  0.68267053  0.39013029
#> 96  096  2.98424301  0.51606200
#> 97  097 -0.94173981  0.88711853
#> 98  098  0.40247013  1.57059197
#> 99  099  0.82147932  1.06667349
#> 100 100  0.39910532 -0.18402181