Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4381254
#> Y_2 0.4381254 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.0229055 -0.03013149
#> 2 -1.6998927 -1.97840525
#> 3  0.7185533  1.00922442
#> 4  0.4735068  0.07745837
#> 5  0.1342741  0.90462880

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.750989692  1.007600940
#> 2   002 -1.085957072 -1.476509507
#> 3   003  0.505740170 -0.600419796
#> 4   004  0.030072106 -0.424924008
#> 5   005  0.756286734  0.532057074
#> 6   006 -0.145121989 -0.285104325
#> 7   007  1.292883582  0.264957934
#> 8   008  1.007646159  0.596250465
#> 9   009 -0.152109136  0.373845010
#> 10  010  0.021832877  0.035004351
#> 11  011 -1.494187756  0.131645600
#> 12  012 -1.121492026 -0.061895703
#> 13  013  1.390094778  1.372795898
#> 14  014  1.082474352 -0.455031467
#> 15  015  0.059892751 -0.018502157
#> 16  016  0.253028953  0.012237463
#> 17  017  0.989186249  0.112136930
#> 18  018  0.789295894  1.820926285
#> 19  019  0.317148265  1.251467858
#> 20  020 -0.541038580 -0.519000976
#> 21  021 -0.375267042 -0.177498846
#> 22  022  0.424541334 -0.232572637
#> 23  023 -0.649499898  0.935386313
#> 24  024  2.328243713  0.800033787
#> 25  025 -0.337699292 -0.707166201
#> 26  026 -0.303936609 -1.104125850
#> 27  027  0.711467229  1.598880222
#> 28  028  1.172678998 -0.332525999
#> 29  029  0.848453643  0.051434081
#> 30  030  0.296184167  0.540822199
#> 31  031  0.569065783  0.851102365
#> 32  032 -0.390260377 -0.019025575
#> 33  033  0.989972775 -0.047912864
#> 34  034 -1.350991163 -2.276455214
#> 35  035  0.878946943 -0.607354540
#> 36  036  0.404298114  0.262526802
#> 37  037  1.593655332  2.046828579
#> 38  038 -0.389263592 -1.452429174
#> 39  039 -1.107638114  0.772633233
#> 40  040  1.280027273 -0.156315379
#> 41  041  0.226330960 -1.088432283
#> 42  042  0.311050944  0.570150486
#> 43  043 -0.508444651 -0.534814493
#> 44  044 -1.533141552 -1.264794867
#> 45  045  0.433354100 -0.659716241
#> 46  046  1.063730385 -0.898384074
#> 47  047  0.341629061  0.490300881
#> 48  048 -0.186439083  0.019655476
#> 49  049 -0.459639719 -0.764068438
#> 50  050 -1.315791399 -0.128927039
#> 51  051 -0.457868963 -1.058209879
#> 52  052  0.648989863  1.195662607
#> 53  053  0.651882960  0.149568661
#> 54  054  0.058194844 -1.865881499
#> 55  055 -0.816229939 -0.216919340
#> 56  056 -0.280585025  0.604121667
#> 57  057 -0.215954747 -1.121278626
#> 58  058  2.328797529  1.532187534
#> 59  059  0.802216890  1.212490075
#> 60  060  1.002595657  1.225066566
#> 61  061  1.395450946  0.120215531
#> 62  062 -0.711336589  0.014641713
#> 63  063  0.575020627 -0.962614448
#> 64  064 -1.336480058 -0.904049635
#> 65  065  0.764306125 -1.094526198
#> 66  066 -0.201178084 -1.226794531
#> 67  067  1.202445814  1.545191092
#> 68  068  0.288346416 -0.448470165
#> 69  069 -0.002403199 -1.543972352
#> 70  070 -0.278106188  0.949630894
#> 71  071  0.549296581 -0.670486570
#> 72  072 -0.104095023 -0.917026765
#> 73  073 -0.504846339 -0.346271054
#> 74  074  1.194608438 -0.167205156
#> 75  075  1.209258169 -0.288805181
#> 76  076 -1.564528163 -1.085336254
#> 77  077  1.731609206  0.987529712
#> 78  078  0.833806362  0.547799431
#> 79  079 -0.253145565 -0.522116025
#> 80  080  0.434880891 -0.091156966
#> 81  081  0.031794108  1.175374068
#> 82  082 -0.862569698 -1.182469440
#> 83  083  0.813803074  0.483198745
#> 84  084 -1.243902179 -0.292149591
#> 85  085 -0.464021518  0.008970747
#> 86  086  1.054830919  0.248231644
#> 87  087 -0.570685501 -1.301089759
#> 88  088  1.000999431 -0.965197362
#> 89  089  0.039203394 -1.413776651
#> 90  090 -0.367386457 -1.327626097
#> 91  091 -1.080001829 -1.211496907
#> 92  092 -0.999983631 -0.813553341
#> 93  093  0.737569339  0.384571952
#> 94  094 -1.352581553 -0.443191024
#> 95  095 -0.380847182  0.016657991
#> 96  096  0.265317519  0.202687151
#> 97  097 -0.265761361 -1.576454132
#> 98  098 -0.263919442  0.165810971
#> 99  099 -0.790482097 -0.522281987
#> 100 100  1.497567511 -0.106206322

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.77192937  1.116392229
#> 2   002  0.08161771  0.479330520
#> 3   003 -0.06079655 -0.168132671
#> 4   004  0.01673018  0.597928408
#> 5   005 -0.75892106 -0.849410736
#> 6   006 -0.77657606  0.583988545
#> 7   007 -0.81899688 -2.069872063
#> 8   008 -1.11270827 -1.002296099
#> 9   009  0.01784684 -0.406704109
#> 10  010 -1.39281874  0.008934189
#> 11  011  1.19681524  1.838962252
#> 12  012 -0.28879737  0.678198848
#> 13  013  0.37183842  1.013602734
#> 14  014 -0.35179213 -0.367665166
#> 15  015 -0.63985868 -0.715085198
#> 16  016  1.42473958 -0.132820331
#> 17  017 -1.29511080 -1.668058710
#> 18  018 -0.22915651  0.348877511
#> 19  019 -0.42218290 -1.790221933
#> 20  020 -1.32570201 -0.502177705
#> 21  021  0.64629326  0.549020944
#> 22  022 -0.60365093 -1.087626671
#> 23  023 -1.10091777  0.695135887
#> 24  024 -1.26749706 -0.969365591
#> 25  025  0.80901142  0.530396578
#> 26  026  1.51902280  2.194742878
#> 27  027 -0.24040571 -0.367668480
#> 28  028 -0.56383202  0.342189518
#> 29  029 -0.46676762  0.354536278
#> 30  030 -1.16605297 -0.888850964
#> 31  031  0.17307159  1.367008362
#> 32  032 -0.39728616  0.278548384
#> 33  033 -0.09675669  1.832387536
#> 34  034  1.96736939  0.376211076
#> 35  035  1.26638544  0.314853495
#> 36  036  0.29229704  0.178455447
#> 37  037 -0.16800731 -1.035887602
#> 38  038 -0.49929012 -0.398865690
#> 39  039  1.13976821 -0.470185119
#> 40  040 -1.45878591  0.032783980
#> 41  041 -0.23231372 -0.989944775
#> 42  042 -0.67710095 -0.044001754
#> 43  043 -0.33618265  0.559800411
#> 44  044 -0.94873658 -0.366941915
#> 45  045 -1.43162752 -0.544439620
#> 46  046 -1.65477131 -2.083407972
#> 47  047  0.30963416  0.225367842
#> 48  048  0.81722571 -0.458206132
#> 49  049  0.46763381 -0.771523468
#> 50  050 -0.29258920  0.321407672
#> 51  051 -1.05664930 -1.178173984
#> 52  052  0.73463865  0.365654289
#> 53  053 -0.62468759  0.107181776
#> 54  054  0.91852106 -0.693839391
#> 55  055 -1.56197041 -0.191286549
#> 56  056  0.77595458 -0.657835147
#> 57  057  0.51885586 -0.710955975
#> 58  058 -0.31429957 -0.959264038
#> 59  059 -1.12317613 -3.105180412
#> 60  060 -1.20193346 -0.810996339
#> 61  061 -0.06433186 -1.651211071
#> 62  062  0.37235188  1.223663918
#> 63  063  0.39816521  0.546173015
#> 64  064  0.75962520 -0.195633018
#> 65  065 -0.42364867  0.435851221
#> 66  066 -1.53627647 -1.720839477
#> 67  067  2.80670586  0.672409671
#> 68  068  0.81662280  0.986921263
#> 69  069 -0.77849590 -0.370544642
#> 70  070 -0.89166505  0.498359526
#> 71  071 -0.97024135 -0.442822088
#> 72  072 -1.75211349 -2.116157210
#> 73  073  0.84016177  0.574347345
#> 74  074  0.50081650  0.818559727
#> 75  075 -0.04881605  0.087872215
#> 76  076  1.54029174  0.806657733
#> 77  077  1.62973566  0.484334218
#> 78  078 -1.31096065 -1.064764411
#> 79  079 -0.08249832  0.600427543
#> 80  080 -0.27607513 -0.144410901
#> 81  081 -0.36262818 -1.018433387
#> 82  082 -1.52043805 -0.392356533
#> 83  083 -0.99146519  0.132777932
#> 84  084 -1.03354988 -1.475097321
#> 85  085 -0.59777428 -1.178874808
#> 86  086 -1.26577282 -0.720411520
#> 87  087  0.17701619  1.150242540
#> 88  088  1.97605104  2.569290043
#> 89  089  0.50238548  1.330608082
#> 90  090  0.34049171  0.892633058
#> 91  091  0.92981794  1.469483389
#> 92  092 -2.00265268 -0.948009404
#> 93  093  0.86922905  0.486624410
#> 94  094  0.89473295  1.683576346
#> 95  095 -0.13358548 -0.038067107
#> 96  096 -0.99107555 -0.943967293
#> 97  097  0.39281075 -0.023638952
#> 98  098  0.50622046  1.794482425
#> 99  099  0.53538652 -1.110473970
#> 100 100  0.04189744  0.377586453