Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5187176
#> Y_2 0.5187176 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.38749366 -1.0394383
#> 2  0.01541053  0.5432468
#> 3 -2.70670488 -2.3759258
#> 4  1.71603286  0.2206089
#> 5  0.73689310  0.2029332

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.13024805  0.802212845
#> 2   002  0.15139058  0.037844917
#> 3   003 -0.21445743 -0.873012908
#> 4   004  1.47661954  1.536024437
#> 5   005  0.66341443  0.919230962
#> 6   006 -0.59064018 -0.368904106
#> 7   007 -1.03334620 -0.140945727
#> 8   008  0.19084154  0.448764158
#> 9   009  1.33716224 -1.100722265
#> 10  010 -0.19759583  0.122979277
#> 11  011 -0.70338898  0.411551880
#> 12  012 -1.32953305 -0.557446330
#> 13  013  1.15992148  0.647399964
#> 14  014  1.16887460  0.264051545
#> 15  015  0.86425535 -0.622046178
#> 16  016  0.33118318  0.223278849
#> 17  017  0.09654428 -0.362873544
#> 18  018  1.25529770 -0.181746604
#> 19  019  0.37559500 -1.264165938
#> 20  020 -0.09970645 -0.504301311
#> 21  021  1.29762480 -0.337463744
#> 22  022  2.63057589  1.824941766
#> 23  023  0.34142480 -0.070399995
#> 24  024 -0.60630684  0.235880870
#> 25  025  0.05073500  0.142307445
#> 26  026  0.06321639 -0.268132177
#> 27  027  0.36830333  0.405625452
#> 28  028 -1.18161982  1.258794179
#> 29  029 -0.26937948  0.586930338
#> 30  030  0.27660858 -1.439760836
#> 31  031 -2.01126642 -0.993237428
#> 32  032  0.22221619 -0.181648526
#> 33  033 -1.98156705 -1.290403984
#> 34  034 -0.51574218 -0.545851413
#> 35  035 -1.78890832 -1.274357432
#> 36  036  0.75035731 -0.345079205
#> 37  037 -0.12631776  0.008093950
#> 38  038  0.28059643  0.287221906
#> 39  039 -2.82142338  1.092948590
#> 40  040 -0.47585920  0.383589073
#> 41  041 -0.13430584  0.110075211
#> 42  042  2.27089217  0.680420675
#> 43  043  1.14253382  0.397865277
#> 44  044 -0.65097498  0.122871925
#> 45  045  0.12204689  0.764155195
#> 46  046 -1.33899590 -1.327286787
#> 47  047 -0.64707140  0.005287322
#> 48  048 -0.54308375 -1.323435656
#> 49  049  0.81441642  0.915690866
#> 50  050  0.13727920  1.110775609
#> 51  051  1.95704293  0.770815323
#> 52  052 -1.23400010 -0.537616744
#> 53  053  0.06135151 -0.176179496
#> 54  054  0.50358147 -1.200004928
#> 55  055 -0.55802459  0.636954358
#> 56  056  0.57687132  2.162846065
#> 57  057 -0.29882998 -0.008891791
#> 58  058  1.98838833  0.981086457
#> 59  059  0.33265175 -0.380829203
#> 60  060 -0.56413554 -0.040282345
#> 61  061  0.30801883  0.432991674
#> 62  062 -0.51216610 -0.384909986
#> 63  063 -1.76198846 -1.567500631
#> 64  064  0.53388116  1.682560501
#> 65  065 -0.75541969 -1.707153524
#> 66  066 -0.19981021 -0.343819028
#> 67  067 -0.07326079 -0.517796598
#> 68  068 -0.31704425  0.482393958
#> 69  069 -2.19532983 -0.299074368
#> 70  070 -1.13185744 -1.085003828
#> 71  071 -0.62598491  0.903252480
#> 72  072 -0.83112606  0.860037957
#> 73  073  1.93661864  1.256594101
#> 74  074 -0.30698529 -1.815612317
#> 75  075  0.51566176  1.817721626
#> 76  076 -0.65337297 -0.431762341
#> 77  077  1.95185608  0.783478603
#> 78  078 -0.48961801 -2.390505238
#> 79  079  0.06995480 -0.128575157
#> 80  080 -0.18815366 -0.887611053
#> 81  081  0.27387431  0.155661378
#> 82  082 -1.80531271 -1.762478946
#> 83  083 -1.28486849 -0.922498182
#> 84  084 -0.36559233 -0.602919078
#> 85  085  0.73021297 -1.145135650
#> 86  086 -1.41155079 -0.713568198
#> 87  087 -0.30014881  0.014465016
#> 88  088 -0.59308730  1.278469849
#> 89  089  1.12029090 -0.082313479
#> 90  090  0.78073850  1.712562854
#> 91  091 -0.73833289 -0.976621864
#> 92  092 -0.38292797 -0.841910010
#> 93  093  0.12543604 -0.011115466
#> 94  094  1.35929644  2.219404955
#> 95  095 -0.75566339 -0.778221171
#> 96  096  2.06005838  1.545755225
#> 97  097  0.35673288 -0.179620653
#> 98  098  0.64004640  0.982353180
#> 99  099  0.94756137  0.184050905
#> 100 100 -0.01101844 -0.188911081

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.12888604  0.21518669
#> 2   002 -0.28006311 -0.35677742
#> 3   003  0.01490609  0.12765846
#> 4   004  0.68494976  0.35742888
#> 5   005 -0.06318544 -1.19997387
#> 6   006  0.41344257 -0.26057902
#> 7   007 -0.96268774 -1.65399450
#> 8   008 -0.34174029 -0.48207895
#> 9   009 -0.08176296  0.88870994
#> 10  010 -0.70007485  0.02807683
#> 11  011  0.41749222  0.99115329
#> 12  012 -0.41392556  1.08366116
#> 13  013 -0.53489862 -0.27984306
#> 14  014  0.17294036 -0.92094291
#> 15  015 -1.28250218  0.75647563
#> 16  016 -1.09350556 -0.03515070
#> 17  017 -1.40837964  0.57256104
#> 18  018  0.38956608  1.96419354
#> 19  019 -0.63729370 -0.31125167
#> 20  020 -0.39497208 -0.84681696
#> 21  021 -0.32846175  0.05079583
#> 22  022 -1.24679282  0.74612139
#> 23  023  0.18863115  0.33621764
#> 24  024  2.70620307  0.53548774
#> 25  025  0.14325811 -0.34014687
#> 26  026  0.64108110  0.48384658
#> 27  027  0.61368719  1.12012989
#> 28  028  1.06634259  0.99598629
#> 29  029 -0.85300510  0.44674160
#> 30  030  1.29095933  0.29115901
#> 31  031 -1.99174891 -1.68918300
#> 32  032  1.92847976  3.01790935
#> 33  033 -0.69030979  1.27617104
#> 34  034  0.53246606  0.86655105
#> 35  035  1.13495171  0.40409828
#> 36  036  0.83771765  1.78310310
#> 37  037  0.28869531  0.37562852
#> 38  038  1.85633815  0.44896116
#> 39  039  0.14056499 -1.34557785
#> 40  040 -1.64887475 -1.74807162
#> 41  041 -0.68297927 -2.17266036
#> 42  042  0.61434672  0.31145543
#> 43  043  0.86819085 -0.28563653
#> 44  044  1.81290661  1.32526749
#> 45  045  0.09753316  1.15124619
#> 46  046  1.05041577 -1.16479390
#> 47  047  0.79553395  1.39412185
#> 48  048 -0.12920346  0.51019001
#> 49  049  2.61473749  2.71716034
#> 50  050 -1.53880713 -0.95260680
#> 51  051 -0.14438692 -0.28535622
#> 52  052 -0.21831631 -0.37685282
#> 53  053  0.17386769  1.31243984
#> 54  054 -1.12948413 -0.99334807
#> 55  055 -1.16924477 -0.28287426
#> 56  056  0.60661710 -0.61968780
#> 57  057 -0.97637595 -2.11284410
#> 58  058 -0.45487693 -0.38390913
#> 59  059 -0.51709756  0.39959195
#> 60  060  1.06175695  3.44758528
#> 61  061  0.11010612  0.32008284
#> 62  062  1.38467871  2.06240414
#> 63  063  0.54340186  0.07729232
#> 64  064  0.95211658 -0.75500263
#> 65  065 -2.20598910 -0.89147885
#> 66  066  0.25163956  0.53565558
#> 67  067 -0.82747776 -0.64425728
#> 68  068 -0.93637596 -1.65553393
#> 69  069  0.86439509  0.65575154
#> 70  070 -0.72842565 -0.88052584
#> 71  071 -0.36515234 -0.50662400
#> 72  072  0.29700921 -0.86488362
#> 73  073 -2.57743300  0.55396394
#> 74  074 -0.29873813  0.53057662
#> 75  075 -1.35819868 -1.15074112
#> 76  076 -0.73318999  0.28500576
#> 77  077  0.37733579  0.30142214
#> 78  078 -0.01010895 -0.53745235
#> 79  079  0.60260638  0.40914863
#> 80  080 -0.30292861 -0.44004362
#> 81  081  1.04345409 -0.53092585
#> 82  082 -0.70535043 -0.11424243
#> 83  083 -0.50491896 -0.91744777
#> 84  084  0.01528792  1.14008376
#> 85  085 -1.05787762 -1.45666583
#> 86  086  1.21566829 -0.83730837
#> 87  087 -0.40584174  0.22425695
#> 88  088  0.92311011 -0.39060270
#> 89  089 -0.37099674 -1.40605684
#> 90  090  0.11604734  0.39070970
#> 91  091  0.44683382  0.31526057
#> 92  092 -1.50494528 -1.32752064
#> 93  093 -0.26429075 -2.27803585
#> 94  094 -1.12089057 -1.77874216
#> 95  095  0.51935127  0.07127268
#> 96  096 -0.97462679 -0.06779422
#> 97  097 -2.07286450 -1.61390071
#> 98  098  0.53579549  0.11216128
#> 99  099  0.83606652  0.53648749
#> 100 100 -0.47174971  0.94741111