Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4882039
#> Y_2 0.4882039 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5701259 -0.3295965
#> 2  1.0628778  0.3122994
#> 3  1.2315535 -0.5114288
#> 4 -1.2082170 -0.7424351
#> 5  0.7856113  1.0390908

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.48764683 -1.03319991
#> 2   002  0.96477114 -0.46868822
#> 3   003  0.19519420 -0.02859214
#> 4   004  0.76527116  0.55637840
#> 5   005 -1.67577338 -1.16337453
#> 6   006  0.33210784 -1.28949230
#> 7   007  0.55900164  0.62296703
#> 8   008  2.41164979  1.68821343
#> 9   009  0.92224879  1.40200680
#> 10  010  0.90453379  0.64871529
#> 11  011  0.27507690 -0.91769499
#> 12  012  1.03933878  0.68632368
#> 13  013  0.89976351 -0.18847184
#> 14  014 -0.40593124  0.28479634
#> 15  015 -0.89162822  1.50839158
#> 16  016 -1.01367505 -1.03132199
#> 17  017 -0.38587961 -1.15621151
#> 18  018  1.44760130 -0.14045895
#> 19  019  0.86759305  0.75269589
#> 20  020 -0.78851676  1.13267675
#> 21  021 -0.35544479 -0.60882561
#> 22  022 -0.26078776 -0.14170318
#> 23  023 -2.46640493 -2.35278003
#> 24  024 -0.17522561  0.51278326
#> 25  025 -0.83289314 -0.27265655
#> 26  026  1.47087968  1.10299593
#> 27  027  0.85918351  0.26781623
#> 28  028 -2.00716765 -0.64593286
#> 29  029 -0.03408010 -0.11483188
#> 30  030 -0.41988422 -0.88185436
#> 31  031 -0.94207438 -2.15805929
#> 32  032  1.55362208 -1.01296303
#> 33  033 -1.64240908 -2.26078623
#> 34  034 -0.06036682  0.07661700
#> 35  035  0.61813131  0.72618712
#> 36  036  0.17006073 -1.21108736
#> 37  037 -1.10583696 -0.84392415
#> 38  038  1.13288807 -0.22093869
#> 39  039 -1.56968713  0.41982073
#> 40  040  1.74074670 -0.39258613
#> 41  041  1.48127376 -0.05233903
#> 42  042  0.92863763 -0.96293595
#> 43  043 -0.51739692  0.83396772
#> 44  044  0.19952121  0.61021555
#> 45  045  0.27297516 -0.42282294
#> 46  046  0.39685064  1.41654411
#> 47  047  0.24077855 -1.78085807
#> 48  048 -1.25776947 -3.14984191
#> 49  049  0.56199121  0.65945816
#> 50  050 -0.62251075  0.13600306
#> 51  051 -0.50206913 -1.00062954
#> 52  052 -0.79517345  0.08949520
#> 53  053  0.05823214 -0.19802063
#> 54  054 -0.25635508 -0.30209544
#> 55  055 -0.16457312  1.00060146
#> 56  056 -0.33514019  0.06157124
#> 57  057 -1.22975222  0.24548823
#> 58  058 -0.06928590 -0.05649609
#> 59  059  0.46954844  0.95329897
#> 60  060 -0.37989411 -0.50949899
#> 61  061 -0.48661803  0.74082462
#> 62  062  1.62676132  0.30203632
#> 63  063 -0.80721961 -1.34415831
#> 64  064 -2.09360348 -0.65636670
#> 65  065 -1.14014109 -1.59056649
#> 66  066  0.60834178 -0.38375255
#> 67  067  0.34818108  0.52192342
#> 68  068 -0.16003507 -0.25291055
#> 69  069  1.55370100  0.97813009
#> 70  070  0.47042744 -0.43453569
#> 71  071  0.52501696  0.28794888
#> 72  072  0.35732242 -0.79519917
#> 73  073 -0.63759958 -0.62450217
#> 74  074  0.70435889 -1.45881947
#> 75  075 -0.07103484  0.39904966
#> 76  076  1.51855901  0.72795270
#> 77  077  0.25214091  0.01397188
#> 78  078  0.55112757 -0.42634814
#> 79  079  1.01930894 -0.68767698
#> 80  080 -0.72837263 -0.95522714
#> 81  081  1.13193885  0.46795041
#> 82  082  0.65498900  1.56405616
#> 83  083  0.78567337  1.15234730
#> 84  084  0.46571712  0.57383496
#> 85  085 -1.69696374 -0.67865954
#> 86  086  0.96498108 -1.11713118
#> 87  087  0.73211825  0.05859363
#> 88  088 -0.52986054 -0.06073311
#> 89  089  2.50354076  2.03799229
#> 90  090 -1.65345203 -0.90518678
#> 91  091  0.26011995 -0.71583971
#> 92  092  0.65166518 -1.67700639
#> 93  093  0.08900399  0.23963716
#> 94  094 -1.46854565 -0.71432531
#> 95  095 -0.08324312 -0.85310378
#> 96  096 -0.65688385 -1.54105946
#> 97  097 -1.85229574 -1.62259296
#> 98  098  0.26055144  0.83597774
#> 99  099 -0.41001970  0.96094502
#> 100 100  0.49749514  1.07000598

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.76237252 -0.009051264
#> 2   002 -1.79029164 -0.341062144
#> 3   003 -0.63896703 -2.900118992
#> 4   004 -0.35228887 -1.308711871
#> 5   005  0.37472316 -0.431857192
#> 6   006  0.36680382 -0.041741917
#> 7   007 -0.30470906  0.178426587
#> 8   008  0.42732002  1.629130957
#> 9   009 -0.94776183  0.330700124
#> 10  010  0.49388710  0.159434426
#> 11  011  0.41647328  0.707457597
#> 12  012  0.64982935 -1.192935648
#> 13  013  0.17176407  0.572674457
#> 14  014 -0.34473561 -0.521242398
#> 15  015  0.08396098  0.475447873
#> 16  016 -1.26169336 -1.476293678
#> 17  017  1.13877907 -0.491368782
#> 18  018  1.18851483  0.566627132
#> 19  019 -1.29458742 -0.251755969
#> 20  020  0.06932386 -1.004158478
#> 21  021  0.13006324 -0.148819968
#> 22  022  1.15732899 -0.118228796
#> 23  023 -1.49754922 -0.434794906
#> 24  024  0.83892369 -0.205209011
#> 25  025  1.24757319  1.743879900
#> 26  026  1.54902861 -0.345360506
#> 27  027 -1.02863434 -0.470547773
#> 28  028 -0.22920750 -0.202009084
#> 29  029  0.12412154 -0.169150173
#> 30  030  2.45682840  2.572538202
#> 31  031  1.11323948  0.451945448
#> 32  032  1.09927092  1.101725817
#> 33  033 -1.11815362  1.381732973
#> 34  034  0.87163700 -0.342305684
#> 35  035  0.20727161 -0.345922425
#> 36  036  0.73347928  0.346885665
#> 37  037 -0.70066847  0.112590358
#> 38  038  0.18833656 -1.018124618
#> 39  039  0.29772407  1.035982584
#> 40  040 -0.28873308 -1.083611797
#> 41  041  1.72482443  0.571919414
#> 42  042 -0.06836532 -2.191388156
#> 43  043 -1.54833080 -0.353384704
#> 44  044 -1.18823068 -0.095263331
#> 45  045 -0.75739034  0.701621287
#> 46  046 -1.44044080 -1.793545623
#> 47  047 -2.35211115 -1.982204384
#> 48  048  0.44176378 -0.003947970
#> 49  049 -0.03265773  0.232154836
#> 50  050 -2.05132129 -1.640661230
#> 51  051 -0.83928085 -1.285099282
#> 52  052  0.68239510 -0.668485195
#> 53  053  1.67417773  0.398803228
#> 54  054 -1.25389983 -0.043011544
#> 55  055  0.09767052  1.714141229
#> 56  056  0.23065644  1.091789321
#> 57  057  0.05326243  1.587042657
#> 58  058  0.43621359  1.168737897
#> 59  059  1.11899636  0.179074366
#> 60  060 -0.41503198 -0.011248049
#> 61  061 -1.70134719 -0.951996772
#> 62  062 -1.08643297 -0.698782433
#> 63  063  0.52580094  0.573669206
#> 64  064  0.61984451  0.216343230
#> 65  065 -1.31833025  0.345826975
#> 66  066  1.29277544 -0.931557007
#> 67  067  0.34431139  1.341293346
#> 68  068 -0.55450627 -0.460588860
#> 69  069 -0.28435683 -0.918295947
#> 70  070 -0.84901931  2.363249045
#> 71  071 -0.94658207  0.667674383
#> 72  072 -0.38404141  0.241843871
#> 73  073 -0.50619059 -1.309874632
#> 74  074  1.86048769 -0.517171507
#> 75  075 -2.64724490 -1.681508486
#> 76  076  1.82938520  0.869554294
#> 77  077 -1.09767584 -0.949414578
#> 78  078  0.75729384  1.389666549
#> 79  079 -1.37758572 -0.536679668
#> 80  080 -0.40843251  0.089938443
#> 81  081 -0.99992504 -0.119435705
#> 82  082  1.10500829 -0.094711236
#> 83  083  0.30436516 -0.576673403
#> 84  084  0.36459389 -0.093773160
#> 85  085 -1.62633119 -0.605916457
#> 86  086  0.79497254  1.029378453
#> 87  087 -0.21072643  0.818609730
#> 88  088  0.36512770  0.577848933
#> 89  089  2.82603760  2.215318117
#> 90  090 -0.67012785 -1.418987886
#> 91  091  1.29140990  1.240017763
#> 92  092 -1.68658697 -0.938845786
#> 93  093  0.33463919 -0.701678857
#> 94  094  2.07404113  2.021526654
#> 95  095 -1.25321035 -1.821533496
#> 96  096  0.04019636 -0.724447980
#> 97  097 -0.80319158  0.405749205
#> 98  098 -1.88483820 -0.692210669
#> 99  099 -1.32047384 -1.240732467
#> 100 100 -0.24096375  0.698989172