Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5380296
#> Y_2 0.5380296 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.78262833 -0.1862132
#> 2  0.05617033 -0.9099106
#> 3 -0.43498299 -1.1824282
#> 4  0.08306032  0.5056838
#> 5 -0.13086905 -1.0908213

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1           Y_2
#> 1   001 -0.33221088 -0.7414417752
#> 2   002  1.49644849  1.0875750212
#> 3   003  1.28297610  0.6295161371
#> 4   004 -1.66786860 -0.8994084045
#> 5   005  1.57586878 -0.3976454861
#> 6   006 -0.68381107 -1.0916153945
#> 7   007  1.15501407  1.1357798064
#> 8   008  0.64421193 -1.3023150385
#> 9   009 -0.20043296  0.8720061536
#> 10  010 -0.76057114  0.3119527536
#> 11  011 -0.21882197 -0.4345328763
#> 12  012 -1.64526360 -0.5852904758
#> 13  013 -0.45052863 -0.2066925166
#> 14  014  0.92533496  1.2883318161
#> 15  015 -0.90713996 -0.9014431949
#> 16  016  0.30834273  1.0137192617
#> 17  017  0.73492670  0.4697790587
#> 18  018 -0.41155383 -0.9911153559
#> 19  019  1.40096715  1.9637949442
#> 20  020  0.57554054  1.7256850074
#> 21  021  0.01327834 -0.2162607667
#> 22  022  0.07726589  0.3524825581
#> 23  023  0.39428279 -0.2461639603
#> 24  024  0.06219582 -0.2476440258
#> 25  025  1.72243010 -0.1774026931
#> 26  026  1.51586483  2.2847001359
#> 27  027 -0.53095234 -0.6252925975
#> 28  028 -0.87498416 -0.4003366020
#> 29  029 -0.35917592 -2.6215608205
#> 30  030 -1.27502207  0.4254418998
#> 31  031 -0.45744631 -0.5644669435
#> 32  032 -1.12760778 -0.4960262814
#> 33  033 -2.39888616 -0.1418522864
#> 34  034 -0.06000971 -0.2288784373
#> 35  035 -0.86812708 -1.6278467129
#> 36  036  2.34018018  1.3689745280
#> 37  037  0.26904237 -0.6089793381
#> 38  038 -0.55665051  0.4244197817
#> 39  039  0.26096177 -0.4507603413
#> 40  040  0.72072585  1.3440822521
#> 41  041  0.07587477 -0.3915524561
#> 42  042 -1.28897767 -0.4496490716
#> 43  043 -0.11244534 -0.8530702120
#> 44  044 -0.06231243  1.7021547936
#> 45  045 -0.67507804 -1.6553755302
#> 46  046 -1.29925585  0.6210671168
#> 47  047  0.28097894  0.8672602082
#> 48  048 -0.26341992 -1.1417808094
#> 49  049  1.44353986  0.9287740690
#> 50  050  0.01833485 -1.8122653198
#> 51  051 -1.89771907 -0.1459432866
#> 52  052  1.30347094  0.0003869112
#> 53  053 -1.58296229 -1.1103130068
#> 54  054 -0.41782554  0.8726877458
#> 55  055  0.92313065  1.3927191507
#> 56  056  0.32546934 -0.2252733789
#> 57  057  0.18209285  1.6743068710
#> 58  058  1.29353154 -0.3528319799
#> 59  059 -1.02501583 -1.5211780223
#> 60  060 -0.61799710 -1.3891023727
#> 61  061 -0.79949472  0.1645561244
#> 62  062 -1.33863972  0.3119490638
#> 63  063  1.43803101  0.3704746144
#> 64  064 -0.52177490 -0.3917805612
#> 65  065 -0.78846752 -0.1279525557
#> 66  066 -0.85738728  0.6863893024
#> 67  067  2.32093306  1.7449529959
#> 68  068  0.52171999 -0.4425804818
#> 69  069  0.94224513 -0.1613622402
#> 70  070  0.79392201 -0.7868842999
#> 71  071 -0.02444483 -0.6665633492
#> 72  072  0.50328264  1.8875631300
#> 73  073  0.15935727  0.6464310533
#> 74  074 -1.52677753 -1.4201666379
#> 75  075 -1.73300253 -0.7275126051
#> 76  076 -1.07717249 -0.3375119063
#> 77  077 -0.23459340  0.3843938055
#> 78  078  1.28182650  0.6367439690
#> 79  079 -0.92395329 -1.2501116047
#> 80  080 -0.71183924 -0.5027359136
#> 81  081  0.49142309 -0.5457241043
#> 82  082 -0.67340547 -1.9560133442
#> 83  083 -0.10327719 -0.0030650593
#> 84  084 -0.45759455 -0.7489548007
#> 85  085  2.21410260  2.0917864319
#> 86  086  1.52415038  1.1732675636
#> 87  087  0.80306980  1.2856077725
#> 88  088 -0.14737225 -0.7072613829
#> 89  089 -1.06212515  0.4691943414
#> 90  090  0.63882932  0.3049350203
#> 91  091  0.47517870  1.2455003188
#> 92  092 -1.85665235 -0.5836588640
#> 93  093  1.25799688  1.2999026970
#> 94  094 -0.66545585  0.0679543638
#> 95  095 -0.47622941 -0.0621871356
#> 96  096  1.84106670  0.3885357464
#> 97  097 -0.40588041 -0.8056995036
#> 98  098  0.01825417  0.1346212195
#> 99  099 -1.50951853  0.2524428465
#> 100 100 -0.21939663  0.0890836707

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.61283623 -1.02827708
#> 2   002 -1.86978287  0.42712239
#> 3   003 -0.77630360  0.46894695
#> 4   004 -0.93234982  0.80414128
#> 5   005  0.55993828 -1.02316815
#> 6   006  1.47687264  1.43495239
#> 7   007  1.23616370  0.67631193
#> 8   008  2.48527128  1.82513134
#> 9   009  0.83060622  1.11796639
#> 10  010  0.87516566 -0.30436633
#> 11  011 -0.29683487 -1.22466503
#> 12  012 -0.49864623  0.63509510
#> 13  013 -0.32963379  0.82975535
#> 14  014  0.52808637 -0.12240684
#> 15  015 -0.04953976  0.78425987
#> 16  016  0.50358926  1.16515919
#> 17  017 -0.57634186 -0.74284327
#> 18  018 -1.15549588 -1.05837852
#> 19  019  0.12979226  0.81859896
#> 20  020  0.91453166 -1.12924479
#> 21  021  1.22059189  1.29234932
#> 22  022 -0.70921727  0.10931447
#> 23  023  2.10219051  1.11573352
#> 24  024 -1.03221492 -1.18639921
#> 25  025  1.68840761  1.15441585
#> 26  026 -0.01059499 -0.13459100
#> 27  027 -1.28377058 -1.29531739
#> 28  028 -0.77098930  0.50053719
#> 29  029 -0.05073303 -0.01272229
#> 30  030 -0.48182436 -0.41141892
#> 31  031  0.92869579  1.61911010
#> 32  032 -1.19029656 -1.94281151
#> 33  033 -0.43342799 -0.28399262
#> 34  034 -1.73850855  0.05651461
#> 35  035 -1.23790650 -0.32869927
#> 36  036 -0.33529162  0.37548796
#> 37  037  0.35980289  0.46997733
#> 38  038 -0.63205447  0.47155514
#> 39  039 -0.10235096  0.02132494
#> 40  040 -0.28408046  1.65297819
#> 41  041  0.78765990  1.96514469
#> 42  042  0.42470364 -0.74136163
#> 43  043  0.64305019 -1.68075018
#> 44  044 -0.82848974 -0.54747600
#> 45  045  0.29784015  0.97556154
#> 46  046 -0.78330247 -0.44982183
#> 47  047 -0.53469460 -1.21901764
#> 48  048  0.30158984  0.52262435
#> 49  049  0.09976344  1.35015918
#> 50  050 -0.67120817 -0.43561879
#> 51  051  2.75140748  2.72438268
#> 52  052 -1.19261623 -1.91050241
#> 53  053 -0.11529681  0.05595876
#> 54  054 -0.10960121  1.30543528
#> 55  055 -0.90960309 -0.21197810
#> 56  056 -1.50970805 -1.45078730
#> 57  057  0.72120072  0.58831775
#> 58  058  0.42140458 -0.36655440
#> 59  059 -1.03034999 -0.29523693
#> 60  060  0.60576106  1.07247804
#> 61  061  0.14445508  0.18992460
#> 62  062 -0.26668218 -0.70733547
#> 63  063  0.37423725  0.49534094
#> 64  064 -0.74641909  0.71595754
#> 65  065 -0.47962878  0.09095007
#> 66  066 -0.13665854 -0.34910984
#> 67  067  0.36178945 -0.10087704
#> 68  068 -0.81732040  1.46727298
#> 69  069 -0.39912906  0.76049596
#> 70  070  0.22815004 -0.44312399
#> 71  071 -0.47162190 -0.66298895
#> 72  072  2.64890105  1.46857748
#> 73  073  0.99967598  0.51115917
#> 74  074 -1.62255518 -0.36228996
#> 75  075  1.27305744  0.12573067
#> 76  076 -1.11033565 -0.45765458
#> 77  077  0.02309524 -0.56314305
#> 78  078 -2.14457649 -0.49932284
#> 79  079 -0.05429208 -1.32841987
#> 80  080  0.61829417  1.75996245
#> 81  081  1.66701593 -0.62634769
#> 82  082 -0.72842964 -1.12575821
#> 83  083  2.06505100  1.46301910
#> 84  084 -0.53515026 -0.48027813
#> 85  085 -0.62075328 -1.20191618
#> 86  086  1.63647894 -0.90010483
#> 87  087 -0.66690286  1.25627237
#> 88  088 -0.42990237 -1.24827213
#> 89  089 -0.49565982  0.64959187
#> 90  090  1.45041489  0.98549965
#> 91  091 -0.09031577 -0.49121188
#> 92  092  0.17879486  0.48148419
#> 93  093 -0.66268971 -0.33617251
#> 94  094 -0.39888700 -0.62827777
#> 95  095 -0.59446885  0.47795339
#> 96  096  0.80295422 -0.91065295
#> 97  097  0.08485752  0.17518790
#> 98  098 -0.58967601 -0.42762169
#> 99  099 -0.74902521 -0.59360750
#> 100 100 -0.50741070 -0.75554468