Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5413237
#> Y_2 0.5413237 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1       Y_2
#> 1  1.5058165 2.3080873
#> 2  0.6794925 0.6847881
#> 3 -0.5599547 0.1005470
#> 4  0.1074610 1.3530211
#> 5 -0.8082354 0.7792588

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.01159525 -0.60646405
#> 2   002  0.65524612  0.84324915
#> 3   003 -0.03454070 -1.24494259
#> 4   004 -1.14432600 -0.68279781
#> 5   005 -0.75126704 -0.58234212
#> 6   006  1.94475986  0.44656689
#> 7   007  0.82357726  0.29062083
#> 8   008 -0.89926493 -2.31698149
#> 9   009  0.78500154 -1.09637801
#> 10  010 -0.75209107 -1.87892504
#> 11  011  0.70854089  0.78729780
#> 12  012  0.28979877 -0.31640451
#> 13  013  0.55926218  0.80646427
#> 14  014  1.14234732  0.62745014
#> 15  015 -0.65979175 -0.44464991
#> 16  016  0.61968651  2.01408466
#> 17  017  0.05917393 -1.05861003
#> 18  018 -0.33407706  0.87934175
#> 19  019  0.75659172 -0.90078539
#> 20  020  0.20057908 -0.16635198
#> 21  021  0.01889023 -0.06817982
#> 22  022 -0.25940420 -1.14894023
#> 23  023 -0.97679860 -0.95388039
#> 24  024  0.66026758 -0.26645898
#> 25  025 -0.02424995  0.72742763
#> 26  026  0.51468822  0.88692701
#> 27  027 -0.70716190 -0.07713598
#> 28  028 -0.61488432 -1.48739007
#> 29  029  1.25126065 -1.06542313
#> 30  030 -0.13346053  0.41699622
#> 31  031 -0.59881870  0.43421607
#> 32  032  1.02588294 -1.33721683
#> 33  033  0.48682454  0.24732711
#> 34  034 -0.18816511 -0.49832143
#> 35  035 -0.51001386 -2.11365702
#> 36  036 -0.20181559  0.85313612
#> 37  037 -0.57438479 -0.11906192
#> 38  038  0.61072024 -0.21195666
#> 39  039  0.45012494  0.27995989
#> 40  040 -1.96097862 -0.07231325
#> 41  041  2.20398845  1.84296832
#> 42  042 -0.84194737  0.09529330
#> 43  043 -1.27964288 -1.90409175
#> 44  044 -0.02531673  0.83262881
#> 45  045 -0.30066394  0.22589479
#> 46  046 -0.02090133 -0.48081277
#> 47  047 -0.67673601  1.50910401
#> 48  048 -1.13563582 -0.71220168
#> 49  049 -0.64364478 -0.45638060
#> 50  050 -0.01266874  0.91241045
#> 51  051 -0.45513198  0.56438657
#> 52  052  2.25451172  0.35196792
#> 53  053  0.45382904  0.78913714
#> 54  054  0.09586737  1.67043274
#> 55  055 -0.80187627 -0.31891416
#> 56  056 -0.10464400 -0.53261337
#> 57  057  1.89092930  1.71597819
#> 58  058  0.64232324 -0.47283749
#> 59  059 -0.24209807 -0.65590252
#> 60  060  1.92529132  0.14814243
#> 61  061 -0.60326952 -0.24867520
#> 62  062 -0.09554963  0.26327136
#> 63  063 -0.02489042  0.11252316
#> 64  064  0.33512725  0.44951447
#> 65  065  0.25462322  0.04231165
#> 66  066 -0.38181773  0.09720750
#> 67  067  1.13609569  1.46259209
#> 68  068  0.96325026  1.02506350
#> 69  069 -0.70137215 -1.69747070
#> 70  070 -1.90198731 -0.39248059
#> 71  071 -0.61855051  0.29214732
#> 72  072  1.33450582  0.72316661
#> 73  073 -0.53430553 -0.45520631
#> 74  074 -2.99073405 -2.58087593
#> 75  075  0.33085790 -0.64738416
#> 76  076 -0.55801763  0.05195660
#> 77  077  0.79738858  0.76814714
#> 78  078  0.09376226  0.31948145
#> 79  079 -0.57210179  1.09753711
#> 80  080 -0.53173318 -1.74562194
#> 81  081 -0.17522191 -0.82876154
#> 82  082 -0.36184405 -0.23063904
#> 83  083  1.20365471 -0.14411034
#> 84  084 -2.46601203 -1.36408970
#> 85  085 -0.79964885 -0.77525042
#> 86  086  1.10325766  1.50885713
#> 87  087  2.46763578  1.35192612
#> 88  088 -0.13660057  0.60483967
#> 89  089  0.05203619  0.07623573
#> 90  090 -0.54061846 -2.12981852
#> 91  091 -0.01626211  0.24005982
#> 92  092  0.01987933  0.84937468
#> 93  093  0.41156083  0.10571101
#> 94  094  0.33613880  2.50119015
#> 95  095 -0.37360833  1.10466374
#> 96  096 -0.18728792  1.21072336
#> 97  097  0.42027476  1.07313864
#> 98  098  1.45635808 -0.26373314
#> 99  099  1.36488896 -0.18550152
#> 100 100  0.03634450  1.85856089

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.808464660  1.57230779
#> 2   002 -0.180505299 -1.87270311
#> 3   003  1.773183725  0.02963279
#> 4   004  0.588921583  0.77896250
#> 5   005 -0.298674517 -1.02757922
#> 6   006  0.703112591 -1.25456381
#> 7   007  0.853322005 -0.26734582
#> 8   008  0.203973887  0.35255661
#> 9   009 -0.645855383 -1.24050695
#> 10  010  1.123554268 -0.38215510
#> 11  011 -0.368689190 -0.08027094
#> 12  012  0.536048947  1.22488468
#> 13  013  0.982636515  0.67964462
#> 14  014  0.305372308  0.60220174
#> 15  015 -0.749114482  1.29333868
#> 16  016  0.232977331  0.52491963
#> 17  017 -0.349386959 -0.70680052
#> 18  018  0.835877105 -0.97227999
#> 19  019 -1.021396811 -0.31963224
#> 20  020 -0.080131526 -0.66024435
#> 21  021 -0.554184770 -0.62035519
#> 22  022  1.204256050  1.00052615
#> 23  023  0.145809893 -0.02764318
#> 24  024  0.225380286  0.08192853
#> 25  025  0.940388047 -0.05058345
#> 26  026 -0.069458659  0.84669419
#> 27  027  0.760510965 -0.05445622
#> 28  028 -0.572271816 -1.13648901
#> 29  029  2.288252095  1.88149135
#> 30  030 -0.209106474 -0.93927945
#> 31  031  0.086060134 -0.31827807
#> 32  032  0.299028834  0.17843661
#> 33  033  1.036413068  0.34466175
#> 34  034 -1.238241773 -0.67959454
#> 35  035  2.626077897  1.36923463
#> 36  036  0.476479484 -0.57673796
#> 37  037  0.011133542  0.22720521
#> 38  038  0.770575304  0.34421090
#> 39  039  1.997482037  0.45253652
#> 40  040 -1.917168051 -0.02565711
#> 41  041 -0.582041921 -0.41360303
#> 42  042 -0.264768718 -1.43448063
#> 43  043 -1.007854352 -0.26377735
#> 44  044 -0.676999593 -1.17224339
#> 45  045  0.030253152  2.16298294
#> 46  046 -0.289104700 -1.76367005
#> 47  047 -0.880776348  0.44468294
#> 48  048  0.485614481  0.10482584
#> 49  049 -1.201895478 -2.31940699
#> 50  050  0.426676224  0.48255536
#> 51  051  0.900082393  0.86780582
#> 52  052 -0.654436340 -0.74048835
#> 53  053  0.237218963 -2.45945994
#> 54  054  0.191700087  0.03683994
#> 55  055 -0.843280346 -1.22461646
#> 56  056 -1.858664882 -0.38992696
#> 57  057 -0.526000063 -1.49060933
#> 58  058 -1.514177816 -1.97488938
#> 59  059 -0.009554014  0.95606373
#> 60  060 -1.255270807 -1.06962046
#> 61  061 -1.122144624 -1.20404402
#> 62  062  1.190324402  0.50422876
#> 63  063  0.368396389 -0.80047515
#> 64  064  0.189965057  0.46077665
#> 65  065  0.903205173  0.87629052
#> 66  066  0.582441719  2.17630347
#> 67  067  2.352319750  3.02892428
#> 68  068 -0.522505749  1.21796942
#> 69  069 -0.198139621  0.10018127
#> 70  070  1.037566256  0.54971072
#> 71  071 -0.187731943 -1.20462635
#> 72  072  0.003232274 -0.40725756
#> 73  073  0.289209282  1.20187569
#> 74  074  0.535435249  2.02454632
#> 75  075  0.707752450  1.09441260
#> 76  076 -0.182522828  0.12250681
#> 77  077  1.122701316  1.79443480
#> 78  078 -1.537971001 -1.99721796
#> 79  079 -0.933092382 -0.83482184
#> 80  080 -1.954867883 -0.88508033
#> 81  081  0.149665909  0.92649801
#> 82  082  0.430225553  0.28457687
#> 83  083 -0.012504403 -0.65131884
#> 84  084  0.112218935 -2.27593348
#> 85  085  0.020376154  0.79895501
#> 86  086  1.226051064 -0.47966352
#> 87  087  2.121910596  1.80145160
#> 88  088 -0.864788243 -0.46301634
#> 89  089  0.400516616  0.52066671
#> 90  090 -0.717955558  0.60697817
#> 91  091  0.308050007 -0.36017217
#> 92  092  1.721438470  0.08438895
#> 93  093 -0.498870197 -1.03026186
#> 94  094 -0.397432745 -1.19225048
#> 95  095  0.215365588 -0.11617980
#> 96  096  0.285428591  0.78154892
#> 97  097  1.284967010  0.19759633
#> 98  098 -1.466201790 -0.56968111
#> 99  099  0.698527937  1.55200730
#> 100 100  0.303522103 -0.24301139