Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>         Y_1     Y_2
#> Y_1 1.00000 0.48773
#> Y_2 0.48773 1.00000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.3396597 -0.4071965
#> 2  0.0706807 -0.6966550
#> 3  1.0193676  0.3472218
#> 4 -1.1073682 -1.4036614
#> 5 -0.2361940  0.7980201

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.930473815 -1.09054210
#> 2   002  0.547146933 -0.74256269
#> 3   003 -0.651555663 -0.06033637
#> 4   004  0.543350711 -0.08477290
#> 5   005  0.718030892  0.97192327
#> 6   006 -0.141190476  0.07572691
#> 7   007  0.798648325  1.73255613
#> 8   008  0.006092835  0.21101318
#> 9   009 -1.063134374 -0.62704932
#> 10  010  0.342753433  0.63192684
#> 11  011  0.370534944  0.32369244
#> 12  012 -1.005074973  1.09270018
#> 13  013  0.929860235 -0.53118822
#> 14  014  0.215086441  0.54262013
#> 15  015  0.194618091 -0.19402944
#> 16  016 -0.346853499 -0.89442234
#> 17  017 -0.673894027  0.15765452
#> 18  018 -0.119380167  0.71265954
#> 19  019 -1.429721161 -0.69606256
#> 20  020  0.709242368 -0.13268538
#> 21  021  0.038069431  0.61531856
#> 22  022 -0.483941363  0.43051971
#> 23  023 -0.035353535  2.09138922
#> 24  024 -1.240591508 -1.37877740
#> 25  025  1.016787797 -0.24165246
#> 26  026  0.704827470 -0.56878920
#> 27  027 -1.006213947 -0.69858198
#> 28  028  0.967683837 -0.70563452
#> 29  029 -0.891154815 -0.99812995
#> 30  030  0.901750843  0.38456406
#> 31  031  0.738973459 -0.31907542
#> 32  032  1.888094204  2.06586486
#> 33  033  0.711185693 -0.08973457
#> 34  034  0.034776898  0.21241037
#> 35  035 -0.391960540  0.26043773
#> 36  036  1.054929389  0.61103203
#> 37  037 -0.916235852 -0.58126464
#> 38  038 -1.240245306 -1.68743490
#> 39  039 -0.689723843  0.61487597
#> 40  040  0.178368279 -0.33025978
#> 41  041 -1.535798238  0.57414795
#> 42  042 -0.080180269 -0.91873620
#> 43  043  0.410733639 -0.39566181
#> 44  044  0.806558523 -0.77269450
#> 45  045 -1.126016293  0.02730397
#> 46  046  0.205307633  0.01209874
#> 47  047  1.853892763  1.49212266
#> 48  048  0.553091725  0.75362523
#> 49  049  0.751036040 -1.32215952
#> 50  050 -1.031541330 -0.92461936
#> 51  051  0.462467405  0.63120563
#> 52  052  0.433264392  0.20220988
#> 53  053  0.681198353 -1.26602210
#> 54  054  0.204668133  0.40924805
#> 55  055 -0.720866327 -0.49065311
#> 56  056  1.663509516 -0.17834660
#> 57  057 -1.175288045 -1.52964803
#> 58  058 -2.449329796 -0.62204549
#> 59  059 -0.217660025 -0.35957182
#> 60  060 -0.618228768 -0.24466187
#> 61  061 -0.547056680 -0.89867338
#> 62  062  0.672276665  1.77028885
#> 63  063  1.403328212 -0.50886592
#> 64  064 -0.547464488 -1.85044731
#> 65  065 -1.263264354 -0.66126490
#> 66  066 -2.324980142 -1.40296493
#> 67  067  2.017247960  2.11384058
#> 68  068 -0.498577189  0.09487993
#> 69  069  1.762383647  1.18582020
#> 70  070  0.312566238  1.41215667
#> 71  071  0.449968784 -0.43893818
#> 72  072 -0.278449427  0.30035597
#> 73  073 -0.521488697  0.48652698
#> 74  074 -0.242790819 -0.24573646
#> 75  075  0.031580802 -0.51696630
#> 76  076 -0.729626586 -2.93486214
#> 77  077  0.064886291 -1.88071044
#> 78  078 -0.230603916 -0.11467502
#> 79  079  1.057982883  0.98048040
#> 80  080 -0.440603819 -0.53931366
#> 81  081 -1.191943801 -0.73515430
#> 82  082 -0.248471228  1.52625909
#> 83  083  0.814755453  0.42698699
#> 84  084 -0.370907112  0.75359218
#> 85  085 -0.736009164  0.35600649
#> 86  086  1.104230394 -0.72193669
#> 87  087  0.719285734  1.47428887
#> 88  088 -0.359629394 -0.04468691
#> 89  089 -0.348842168  0.11344686
#> 90  090 -0.029680541 -1.04642692
#> 91  091  0.741860527  1.81761267
#> 92  092 -0.553809846  0.09078332
#> 93  093 -0.187869930 -0.32406179
#> 94  094 -0.258872572  0.13759027
#> 95  095 -1.704157034 -1.92106236
#> 96  096  1.549128266  1.60498757
#> 97  097  0.688947293  1.48992032
#> 98  098  0.829400045 -1.20843412
#> 99  099 -1.659954493 -0.92927447
#> 100 100  1.682779637  0.85252387

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -1.08145143 -0.5864188257
#> 2   002  0.51121925 -0.3029995360
#> 3   003 -1.24117446 -0.6489114552
#> 4   004  0.95407837  1.4106134620
#> 5   005 -2.37837394 -1.0959204804
#> 6   006  0.55774932 -0.6165759269
#> 7   007 -0.63474839 -0.0002882232
#> 8   008  0.60369409  0.2068827208
#> 9   009 -1.71468916 -0.6194425878
#> 10  010 -2.35372711 -2.8846854598
#> 11  011 -1.56577169  0.1145377756
#> 12  012 -1.07380380 -0.8218182062
#> 13  013  0.83514287  0.1441703357
#> 14  014 -0.96052896 -0.1829294460
#> 15  015 -0.46637397  0.2647109934
#> 16  016 -0.70800112 -0.0053819171
#> 17  017  0.09410629 -1.1672520314
#> 18  018 -1.18028037 -0.7234778946
#> 19  019  1.17996824  0.7510229695
#> 20  020 -0.92681693 -2.0350174748
#> 21  021 -1.02573574 -1.2444996781
#> 22  022  0.44224344 -0.4772151627
#> 23  023  1.22904104  1.0169634427
#> 24  024  0.96221525  0.7812340833
#> 25  025  0.04357757  0.0039473312
#> 26  026 -0.21505560  0.8209932703
#> 27  027  1.23052216 -0.4334055669
#> 28  028  2.42516222  0.5637634191
#> 29  029 -1.40886678 -1.2587866791
#> 30  030  1.70727566  1.2399633360
#> 31  031 -0.04492504 -1.2283205718
#> 32  032  0.49349167  1.0151433394
#> 33  033 -0.65742282 -0.6428230348
#> 34  034 -1.62716800  0.1295808006
#> 35  035 -0.51260677 -1.0424398646
#> 36  036 -0.49313634  0.1724352906
#> 37  037  0.02306197  0.5131136224
#> 38  038 -1.87954979 -1.4654354755
#> 39  039 -2.07244926 -1.6944851065
#> 40  040  0.82736338  1.1239128888
#> 41  041 -2.00888526 -1.0401281109
#> 42  042 -0.28939439 -0.7161319277
#> 43  043 -0.80818571 -0.0820386615
#> 44  044 -0.56368744 -1.3937274642
#> 45  045 -0.73970589 -1.3040425976
#> 46  046 -0.06341147 -0.0807411430
#> 47  047 -0.35801613  0.1395602603
#> 48  048  0.66014244  0.8563911235
#> 49  049  1.60078292  1.5344993114
#> 50  050  1.72239187  0.0258739558
#> 51  051  0.34673082  1.6885054623
#> 52  052  0.06997432 -0.2944680121
#> 53  053 -0.23139057  0.1278580171
#> 54  054 -0.08934846 -1.0465141836
#> 55  055 -1.13664944  0.0942340402
#> 56  056  0.46992118 -0.4334629180
#> 57  057  1.30331287  0.6614059894
#> 58  058 -1.08959360  0.3499253808
#> 59  059 -2.19293726 -1.4430000945
#> 60  060 -0.62038328  0.1974075255
#> 61  061  1.81812148  0.5767373998
#> 62  062 -1.59978140 -0.2352627111
#> 63  063 -1.49397270 -0.6981895115
#> 64  064  0.75375289 -1.1764347269
#> 65  065  0.78247194  1.6628201330
#> 66  066  0.53343932  1.2431138246
#> 67  067 -0.18666209 -0.5363650206
#> 68  068  0.06137770  0.0403887574
#> 69  069 -1.62111510 -1.8195247651
#> 70  070  0.69204439 -0.8338980137
#> 71  071  0.51677714 -0.6389279416
#> 72  072 -0.55940419  0.0779211115
#> 73  073  0.58261368 -0.6698770265
#> 74  074 -0.78660818 -0.5065585195
#> 75  075 -1.46064853 -1.8868750835
#> 76  076  1.55249575 -0.5938107773
#> 77  077 -0.74176940  0.7315402974
#> 78  078  0.70725534  1.6735076152
#> 79  079  0.27871318  0.0216156054
#> 80  080  0.77743407  1.3917754324
#> 81  081  0.37623392  1.0189918335
#> 82  082  0.86695378 -0.2331989048
#> 83  083 -0.29392314 -2.5087110328
#> 84  084  0.95547094  1.0197211565
#> 85  085 -2.28687019 -1.2090491185
#> 86  086  0.49011924 -1.0927034028
#> 87  087 -0.46664962 -1.1232539944
#> 88  088  0.12531107 -0.3758089743
#> 89  089  0.04600169  0.4228587951
#> 90  090  0.51553335  0.2895674099
#> 91  091 -0.49684737  0.8795762689
#> 92  092 -1.88037826 -1.2331499477
#> 93  093  0.12181396 -0.4567484914
#> 94  094  0.84012816  1.7132590281
#> 95  095  0.66798496  0.1873333553
#> 96  096 -0.65473778  0.0574367861
#> 97  097  0.93604519 -0.0896802784
#> 98  098  0.13878948  0.1928821944
#> 99  099 -0.54718572 -0.1927816084
#> 100 100 -1.36498882 -1.2823296503