Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>         Y_1     Y_2
#> Y_1 1.00000 0.43189
#> Y_2 0.43189 1.00000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  1.3628242  2.12603401
#> 2  0.8114644  1.57890580
#> 3  1.9497901 -0.28817998
#> 4 -1.3808907 -1.26082703
#> 5 -1.4973733 -0.07460092

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.582521795 -2.030101776
#> 2   002 -0.056414067 -0.203245988
#> 3   003  0.951888489  0.039635857
#> 4   004  1.248714909 -0.033741282
#> 5   005  1.211802859 -1.163126289
#> 6   006  0.647615220  1.059548195
#> 7   007  0.333703966  0.072419584
#> 8   008  0.695301324  0.268397811
#> 9   009 -0.174405158 -0.162869740
#> 10  010 -1.011444492  0.308348048
#> 11  011  0.357905331 -0.525439495
#> 12  012 -2.247485391 -0.723348187
#> 13  013 -3.047617976 -1.793215086
#> 14  014  1.643557796  0.914026816
#> 15  015  0.510794428 -1.147104285
#> 16  016  3.114903957 -1.043791471
#> 17  017 -0.531238114 -0.040130834
#> 18  018  0.172113953  0.029238952
#> 19  019 -0.098991915 -1.035802941
#> 20  020 -0.699217629 -0.313280217
#> 21  021 -0.090350215 -0.687451555
#> 22  022 -0.006624928  0.072389163
#> 23  023  0.593712826  0.695841344
#> 24  024  1.117575586 -0.946876056
#> 25  025  0.821788882 -0.327098118
#> 26  026  0.242784340  0.288734513
#> 27  027 -0.812900021 -1.463561385
#> 28  028 -0.849685012 -0.369705539
#> 29  029 -0.547156832  1.471155169
#> 30  030  0.084426218  0.182581280
#> 31  031  0.492650557 -1.543676494
#> 32  032 -0.647554019  0.344305341
#> 33  033 -0.697549286 -1.866637286
#> 34  034 -0.059846178 -0.413880219
#> 35  035  0.179251002  1.151855066
#> 36  036 -0.078162607  0.663166540
#> 37  037  1.088352713  0.300030512
#> 38  038  0.442937440  0.613152501
#> 39  039 -0.795973659 -2.219902341
#> 40  040  0.149583524  1.903900401
#> 41  041 -0.766092464 -1.389255412
#> 42  042  2.497455276  0.532416829
#> 43  043 -0.868756462 -2.269286099
#> 44  044 -0.867995003  0.176169818
#> 45  045  0.441888506  0.443823992
#> 46  046 -1.377383672 -0.428044289
#> 47  047 -0.482772052  0.651253480
#> 48  048  0.046103294 -0.099418536
#> 49  049  0.638101345  1.167294233
#> 50  050  1.003068681 -0.689053218
#> 51  051 -0.069978753 -0.001492666
#> 52  052  1.408120567  0.383501221
#> 53  053  1.440507128 -0.283380143
#> 54  054  0.331015325 -0.218363856
#> 55  055  0.823489591  1.021711158
#> 56  056  1.685426892  0.224869729
#> 57  057  1.043095212  1.323442711
#> 58  058 -0.324223821  0.357606476
#> 59  059 -0.086482742 -0.053365084
#> 60  060  0.027490537  0.381161478
#> 61  061  0.465417103  1.398047739
#> 62  062  1.199543596  2.532240404
#> 63  063  1.216530305  0.085613533
#> 64  064  1.214580268  0.963577262
#> 65  065  1.074438778  0.227982387
#> 66  066 -0.036083965  0.714542695
#> 67  067 -0.499539557 -0.204615136
#> 68  068  1.043909761  2.684144869
#> 69  069  0.339103946  0.348011011
#> 70  070  0.912747504  0.212258087
#> 71  071  0.614062728  0.921231925
#> 72  072 -0.111738129 -0.750412815
#> 73  073 -0.406856410  0.791699199
#> 74  074  0.593929731  1.616133139
#> 75  075  1.208289383 -0.059742541
#> 76  076  0.540492620  0.166266055
#> 77  077  0.074085273 -0.769428986
#> 78  078  0.360242800  0.734030443
#> 79  079 -0.324890377  0.137300539
#> 80  080 -1.352819225 -1.007750939
#> 81  081 -1.613530289 -0.041118467
#> 82  082 -1.619611145 -2.093006155
#> 83  083 -1.047497233  0.131594241
#> 84  084  0.305445179 -0.757254270
#> 85  085 -0.248831212 -0.804935194
#> 86  086 -1.640872758 -1.484263258
#> 87  087 -0.001505957 -1.759986126
#> 88  088 -0.033342804 -0.891355584
#> 89  089  1.185815312  0.565391045
#> 90  090 -1.735788518 -1.857004792
#> 91  091 -0.876648265 -0.840041838
#> 92  092  1.037428738 -0.057939246
#> 93  093 -0.400791253  0.828536150
#> 94  094 -0.329303778  0.945884547
#> 95  095  0.564961855  1.880612811
#> 96  096 -1.079107189  0.291404387
#> 97  097  1.295516965  0.820421082
#> 98  098  0.306873617  1.373027291
#> 99  099  0.697016402 -1.256786609
#> 100 100 -0.779800477  0.590702260

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.772562941  2.12798672
#> 2   002  1.003919904 -1.23630619
#> 3   003 -0.937429327 -0.67726475
#> 4   004 -0.475288463 -0.24906003
#> 5   005  1.582922150  1.12176493
#> 6   006 -1.012488178 -1.15764716
#> 7   007 -0.407445454  0.44425621
#> 8   008  0.025399197 -0.62127853
#> 9   009 -0.003501335 -0.40951301
#> 10  010  0.260613367  0.11512892
#> 11  011 -1.982475570 -0.35000076
#> 12  012 -0.441406315 -0.17389452
#> 13  013 -1.315973699 -1.14587000
#> 14  014  1.237204916  0.41278527
#> 15  015  1.229518459  1.24560920
#> 16  016 -1.490021298  1.36003159
#> 17  017 -1.820748876 -3.05402365
#> 18  018  1.179220582  0.90018996
#> 19  019 -1.028897032  0.59456346
#> 20  020 -0.357077195 -0.65671265
#> 21  021  1.338253907  0.33704522
#> 22  022 -0.034547707 -0.94611809
#> 23  023 -0.540961279 -0.54477888
#> 24  024 -0.174971256  0.48835166
#> 25  025  0.024720818  0.58967366
#> 26  026  0.824943049  0.86447845
#> 27  027  1.541159182  0.62997481
#> 28  028  0.341995992 -0.19803686
#> 29  029  0.539632422 -0.53480336
#> 30  030  0.931091110  1.10195253
#> 31  031 -0.597389337  0.05922378
#> 32  032 -1.146781834  0.50127143
#> 33  033  1.245674164 -1.45735092
#> 34  034 -1.091482883 -0.08000106
#> 35  035 -0.998958160 -1.60925514
#> 36  036  0.942283590  0.77305538
#> 37  037 -0.529507598  0.66341881
#> 38  038 -0.142016706  1.16127007
#> 39  039  0.404856824  2.15067093
#> 40  040 -0.131249919 -0.20569657
#> 41  041 -1.078065688  0.06252075
#> 42  042  1.104313204  1.91175404
#> 43  043  1.546527259  0.94322364
#> 44  044 -0.192854689  0.77976845
#> 45  045  1.261759039  3.07063819
#> 46  046  0.089204122 -0.36118404
#> 47  047 -0.721841380  1.16059012
#> 48  048  1.438264137 -0.37099237
#> 49  049  0.943872609  3.34768672
#> 50  050  0.099406353 -0.11134809
#> 51  051 -0.545165995  0.06710034
#> 52  052  0.872739658  0.10874778
#> 53  053  1.332571323  0.77976343
#> 54  054 -2.164582022 -1.71930480
#> 55  055 -0.634956654 -0.10517725
#> 56  056  0.634187190  0.70593131
#> 57  057 -1.179210544 -0.36879964
#> 58  058  0.731982000  0.24987534
#> 59  059  0.770919970 -0.86559489
#> 60  060 -1.345195465 -0.91508195
#> 61  061  0.180387639 -0.33376240
#> 62  062  0.794829581 -0.53371915
#> 63  063  0.191610057 -0.22513839
#> 64  064 -0.275290657  0.12565185
#> 65  065 -0.429202331  0.15682697
#> 66  066 -0.291394684 -0.70025843
#> 67  067  0.506025544 -0.28973092
#> 68  068  1.103140312 -0.49692280
#> 69  069  0.071046277 -0.92448673
#> 70  070  0.803051310  0.43618762
#> 71  071 -0.327880749 -0.75452369
#> 72  072  0.404815919 -0.24676413
#> 73  073  1.932468227  1.21758789
#> 74  074 -0.261387251  0.46230067
#> 75  075  0.487200187 -0.32797564
#> 76  076 -0.677820902 -1.94371487
#> 77  077 -0.115867791 -0.89320168
#> 78  078 -1.148788488 -1.52422111
#> 79  079  0.143033534 -1.54967709
#> 80  080 -0.541740454 -0.39272589
#> 81  081  0.802557450  1.32521428
#> 82  082 -0.012167203  0.77789556
#> 83  083 -0.122188868  1.11418021
#> 84  084  1.567317463  1.08155868
#> 85  085  1.789266628 -0.19017944
#> 86  086 -1.016211529  0.68186394
#> 87  087  0.593509659 -0.05307833
#> 88  088  0.520675350  0.75658436
#> 89  089  0.421329259  0.88882381
#> 90  090 -0.088235577 -0.45326116
#> 91  091 -0.782625313 -0.81055033
#> 92  092 -1.351681206  0.35916577
#> 93  093  1.169763503  1.10329965
#> 94  094 -0.168398196 -0.52839275
#> 95  095  0.148447844  0.52572098
#> 96  096 -1.138051402  0.36025728
#> 97  097 -0.131761288  1.18986636
#> 98  098 -0.573052952  0.24249017
#> 99  099 -0.024392680 -0.47621927
#> 100 100  0.547020076  0.10657498