Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.439654
#> Y_2 0.439654 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.2684167  0.82270995
#> 2 -0.2961757  0.01204274
#> 3 -0.3238774  0.42190965
#> 4  0.3553881 -0.55592145
#> 5 -0.3767354  1.45831438

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.298293465  0.522488383
#> 2   002  0.513381639  0.000502089
#> 3   003  0.584514713  1.979323926
#> 4   004  0.266541631  0.011725353
#> 5   005  0.909360773  1.525432965
#> 6   006 -1.459386778 -1.477796246
#> 7   007  0.897642192 -0.151628594
#> 8   008 -1.149129008  0.262378797
#> 9   009 -0.712226251 -0.927687407
#> 10  010 -1.000747824 -1.098343753
#> 11  011 -0.027097733 -0.747908765
#> 12  012 -0.288676463 -0.158490232
#> 13  013  0.349297982  1.717697341
#> 14  014  0.696630611  0.317870157
#> 15  015  0.044751754 -0.845756688
#> 16  016  0.356755019 -2.109080894
#> 17  017  0.769628644  1.933045861
#> 18  018 -0.953924353 -0.711748265
#> 19  019 -0.543154240 -0.439405162
#> 20  020 -0.362492249 -0.059778567
#> 21  021 -1.859062704 -1.350396234
#> 22  022 -0.488219936 -0.025496271
#> 23  023  0.762782171 -0.167216685
#> 24  024 -1.286853716 -1.435459307
#> 25  025  0.047970630 -0.874738499
#> 26  026  1.541369013  1.121325982
#> 27  027  1.970914062  1.154656549
#> 28  028 -0.878073779 -0.693543902
#> 29  029  0.787500239  1.129242383
#> 30  030  1.116026799 -0.356163527
#> 31  031  0.345665602  1.232849073
#> 32  032 -0.027167260  0.171467332
#> 33  033  0.278426968  0.125387952
#> 34  034  0.489955660 -0.939597975
#> 35  035 -0.835363629  0.832764281
#> 36  036  1.898396572  0.374431303
#> 37  037  1.239454005 -0.236393134
#> 38  038  0.721669667  0.987892362
#> 39  039  0.445635161 -1.679433030
#> 40  040  0.829749704  0.057744485
#> 41  041 -0.129367754 -0.071916128
#> 42  042 -0.482805346 -0.263734678
#> 43  043 -0.579786745 -0.319063943
#> 44  044  1.065909506  1.237993980
#> 45  045  0.123257167 -0.341415072
#> 46  046 -0.066347193  0.457898652
#> 47  047 -1.112565819 -0.552856831
#> 48  048  0.524967942  0.274398833
#> 49  049 -1.077760784 -0.634229050
#> 50  050  0.420588123 -1.205196298
#> 51  051 -1.695995900 -1.005862534
#> 52  052  0.672712676  0.894313747
#> 53  053  1.150684082 -0.350721946
#> 54  054  0.699745799  0.160662860
#> 55  055 -0.588094613 -0.243849081
#> 56  056  0.615543316  0.327130088
#> 57  057  1.171424226 -0.512038856
#> 58  058 -0.462288197 -0.328052867
#> 59  059 -1.482913633 -1.541003567
#> 60  060 -1.317463904  0.506703935
#> 61  061  1.237248106  1.032001302
#> 62  062  0.383647469  1.073664477
#> 63  063 -1.029342797  1.294552334
#> 64  064 -0.076078420  0.086122312
#> 65  065  0.446617354 -0.311570905
#> 66  066 -0.624169888  0.090657117
#> 67  067  1.479656180  1.818921610
#> 68  068 -1.459328439 -1.460601999
#> 69  069  0.921703987  0.133716101
#> 70  070 -1.058612788 -1.069043001
#> 71  071 -0.142933311 -0.553280718
#> 72  072  0.712919808  1.365949162
#> 73  073  2.948790790  2.436868620
#> 74  074 -0.834965806 -1.208971641
#> 75  075  1.029807283 -0.309268023
#> 76  076 -0.309424688  1.025057912
#> 77  077  1.208253705  2.005450398
#> 78  078  0.015824362  0.035126323
#> 79  079  0.315072006  2.682613983
#> 80  080  0.634080975  0.567953077
#> 81  081  0.013898806  0.375424746
#> 82  082 -0.020059914  0.197716659
#> 83  083  0.556694234  1.170514502
#> 84  084 -0.136711829  0.139839785
#> 85  085 -0.005717582 -0.220540225
#> 86  086  1.698329734  0.688506580
#> 87  087 -0.761608469  0.407336267
#> 88  088  1.073526135  0.276401971
#> 89  089 -1.595272910 -1.356205370
#> 90  090 -0.500266281  0.424297012
#> 91  091 -0.607799012  1.437515324
#> 92  092 -1.064111279 -0.521247900
#> 93  093  2.544336492  1.034998754
#> 94  094  0.801361985 -1.241891604
#> 95  095 -1.850746855 -0.583183288
#> 96  096  0.743951261  0.845137765
#> 97  097  0.578297158  1.006375110
#> 98  098  0.488708178  0.125660025
#> 99  099  0.234676977  0.627804859
#> 100 100  1.206145783  1.207852054

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1          Y.2
#> 1   001  0.1566406636 -1.127825562
#> 2   002  0.5234975489  0.502292421
#> 3   003 -0.8641270506 -0.521869010
#> 4   004  0.1242292207  2.478090728
#> 5   005 -0.0953827072  0.047477521
#> 6   006 -0.4008517045  0.503061670
#> 7   007 -1.6057516293 -0.311285228
#> 8   008 -0.4975984250  1.387190709
#> 9   009 -0.7336778625 -0.436556674
#> 10  010  0.2581317541  2.293118401
#> 11  011  0.3703948621  0.626121455
#> 12  012  1.1073495839  1.399302154
#> 13  013  0.8608208140 -0.438168781
#> 14  014 -1.2502887974  0.594406816
#> 15  015  0.5053908932  0.449100554
#> 16  016 -0.4990823110 -1.005542200
#> 17  017  0.6117540116  1.455979086
#> 18  018 -0.8860631382  0.665661417
#> 19  019  0.4473721619  1.178646716
#> 20  020 -2.6206836867 -1.809164468
#> 21  021  0.0526838978 -0.082449060
#> 22  022  0.4866110109  0.778255950
#> 23  023 -0.4623647854 -1.074132216
#> 24  024  0.2427277239 -0.828918699
#> 25  025 -0.0152171333  0.539982814
#> 26  026  0.8845729651 -1.956783244
#> 27  027 -1.2926326927  0.128534148
#> 28  028  0.3629144428  0.973240547
#> 29  029  0.2931375694 -0.373511804
#> 30  030  0.7983019715  0.083647983
#> 31  031 -0.2890729356 -0.549468633
#> 32  032 -0.2969391080  0.031280579
#> 33  033 -0.3244352265 -0.476122240
#> 34  034  1.0607225074  0.722136047
#> 35  035 -0.6803013777 -0.531239179
#> 36  036  0.0886432635  0.408330572
#> 37  037 -0.9103546815 -1.351590822
#> 38  038  0.3721531534  0.822225882
#> 39  039 -0.5721363818 -0.290086092
#> 40  040 -1.1496903774 -0.150031336
#> 41  041 -0.4926163867 -0.244580529
#> 42  042 -0.9965702673 -0.846992818
#> 43  043 -1.9630219231 -2.216503620
#> 44  044  0.2258656707  0.968267933
#> 45  045 -0.5177000974 -0.896563718
#> 46  046  0.6553336038  0.951733090
#> 47  047 -0.6277837459 -1.115937997
#> 48  048  0.6947291277  0.554206917
#> 49  049 -0.4701880658 -0.647142726
#> 50  050 -0.2384741477 -1.009630005
#> 51  051 -0.4103260864  0.172373169
#> 52  052 -0.2174737952 -0.268519111
#> 53  053  0.0114147381  0.771734276
#> 54  054 -0.7760096467 -1.063849755
#> 55  055  0.3904258290 -0.465472517
#> 56  056  0.5733058204 -0.175488024
#> 57  057 -0.4150869127 -0.744387979
#> 58  058 -0.4336606144 -0.445036483
#> 59  059  0.2940972658 -0.336296245
#> 60  060 -0.9754809192 -1.545292393
#> 61  061  1.8164485445 -1.163053674
#> 62  062  0.6772644415  0.094621575
#> 63  063 -0.3482267888 -0.594766477
#> 64  064 -2.0609357899 -1.486082846
#> 65  065  0.2099495242  1.208400237
#> 66  066  3.5810560993  1.757219229
#> 67  067  0.0686660223  0.708092475
#> 68  068 -2.0575616726 -0.147075840
#> 69  069  1.7050968747  0.480031891
#> 70  070 -0.1095405840 -0.884135973
#> 71  071  0.7953422453  0.012390041
#> 72  072  0.4126986046  0.351633919
#> 73  073  0.3997480873 -0.238875312
#> 74  074 -0.9088624788 -1.456822815
#> 75  075 -0.9631089972 -1.143957948
#> 76  076  0.9914692198 -0.593533120
#> 77  077  0.1259306641  0.547211988
#> 78  078 -0.6283897557  0.692590643
#> 79  079 -0.3838066628 -0.178045307
#> 80  080  0.3608410450  0.529868534
#> 81  081  0.4716863032 -0.226293584
#> 82  082  0.0235492645  0.005710703
#> 83  083  0.2493782011  0.420632983
#> 84  084 -0.9587881249 -1.415657758
#> 85  085 -0.7380179039 -0.622722160
#> 86  086 -2.2812454350 -2.337418294
#> 87  087 -0.0003725389  0.289117476
#> 88  088 -0.3503981790  1.082412624
#> 89  089  0.0552039373  1.332284651
#> 90  090  0.7965329773  3.281709529
#> 91  091  0.4662769072 -1.267503802
#> 92  092 -0.0848889863  0.472458761
#> 93  093  1.7535755931  0.599079073
#> 94  094 -2.2978313601 -0.981264771
#> 95  095 -1.8517982204  0.117551989
#> 96  096 -0.0634831029  0.187746865
#> 97  097 -0.4614131676  1.070233565
#> 98  098  0.9769445141 -1.553537021
#> 99  099  1.2025037399  0.389008747
#> 100 100  0.4880227362 -1.152200592