Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5299487
#> Y_2 0.5299487 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.7850553 -2.6261939
#> 2  1.3380191  0.4627640
#> 3 -0.5187482 -0.5932778
#> 4 -0.2029016 -0.7780516
#> 5 -0.8225381 -0.6891169

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.691448559  0.211242729
#> 2   002 -0.997151624 -1.900435337
#> 3   003 -0.301151582  0.678403142
#> 4   004  1.041105579  1.127488975
#> 5   005 -0.942455255 -0.245404441
#> 6   006 -0.697865505  0.546409504
#> 7   007  1.275985509  0.689647140
#> 8   008  0.206193570 -0.588758747
#> 9   009 -0.951838350 -1.596393959
#> 10  010 -0.682279665  0.536389968
#> 11  011 -0.323697661 -0.061884479
#> 12  012 -1.067835873 -0.746449012
#> 13  013  0.002143816  0.648700411
#> 14  014  1.324436615  0.978842023
#> 15  015  0.040425813  0.177678486
#> 16  016 -0.178452824 -0.303872652
#> 17  017  0.797032443 -0.907492818
#> 18  018 -0.920323992  0.331382141
#> 19  019  0.054170285 -0.880066609
#> 20  020 -0.391663852 -2.579029108
#> 21  021  1.637008494  0.693483593
#> 22  022  0.815777017  0.231706416
#> 23  023  2.249361318  1.270503313
#> 24  024  1.339677815  1.776944834
#> 25  025  1.567058128  0.332672451
#> 26  026  0.058489068 -0.156391352
#> 27  027 -0.013779919 -2.095338809
#> 28  028  0.504324276  1.245475576
#> 29  029 -0.930798554 -0.639972992
#> 30  030  0.097475491  0.193920467
#> 31  031 -0.787141096  0.057174124
#> 32  032 -0.378490188 -1.734587151
#> 33  033 -1.354618301 -0.381433370
#> 34  034  0.232630299  1.516928860
#> 35  035 -1.416532257  0.375310104
#> 36  036 -0.289149007  0.488334493
#> 37  037 -0.042175609  1.580856980
#> 38  038 -1.170903258  0.023746453
#> 39  039  1.086080198  1.394393701
#> 40  040 -1.146150388  0.573697609
#> 41  041 -3.651480198 -1.902260429
#> 42  042  0.535035927  0.295432390
#> 43  043 -0.469243154 -0.735278951
#> 44  044 -1.084787060 -2.036289739
#> 45  045 -2.325074903 -2.389366281
#> 46  046  0.900075306  0.150888649
#> 47  047 -0.134327358 -0.506044736
#> 48  048  0.461031729 -0.721018516
#> 49  049 -0.714835631  0.592802728
#> 50  050  0.787475184 -1.148397418
#> 51  051 -1.276761632 -0.416349746
#> 52  052 -0.270029216  0.505391624
#> 53  053 -1.155153664 -1.033865033
#> 54  054  0.080337782 -0.793387853
#> 55  055 -0.305937255 -0.588759896
#> 56  056 -0.249246246  0.484598182
#> 57  057  0.704249435  1.855138053
#> 58  058 -1.196925351 -0.625264521
#> 59  059  1.433162694  0.885109683
#> 60  060  1.249271975  1.086197515
#> 61  061  0.607461345  0.710458501
#> 62  062 -0.730775333 -1.017346847
#> 63  063  0.360964771  1.344440093
#> 64  064  0.138444321  0.096839142
#> 65  065  0.414217657 -1.129581019
#> 66  066  0.770267758  0.150023500
#> 67  067 -0.618697265 -1.235457064
#> 68  068 -1.715749461  0.501040337
#> 69  069 -1.158970144 -2.535246272
#> 70  070  0.982479074  1.841130820
#> 71  071  1.756914892  1.525622509
#> 72  072 -0.569172966  0.105511110
#> 73  073  0.124924272 -1.173573873
#> 74  074  0.511940934 -0.208439745
#> 75  075 -1.509300845 -0.055518842
#> 76  076 -1.260015299  0.125852046
#> 77  077 -1.184945066  0.519702817
#> 78  078 -0.886222852 -0.629139289
#> 79  079 -1.423214655 -0.441152542
#> 80  080  0.810268207  0.059225340
#> 81  081  0.279664581  0.002086859
#> 82  082 -0.067172021 -0.682477790
#> 83  083 -0.163610940  1.868654381
#> 84  084 -0.437606660  0.669222199
#> 85  085 -0.763596858 -1.211362666
#> 86  086 -0.600325133 -0.630741254
#> 87  087  1.172658947  3.045286461
#> 88  088 -0.256534354 -0.573764548
#> 89  089  0.842164073 -0.033399117
#> 90  090  1.612385912 -0.905342852
#> 91  091  0.546624225 -2.120111514
#> 92  092  1.725632546  2.162086765
#> 93  093 -1.291579594 -1.011751508
#> 94  094 -0.351260374  0.699400572
#> 95  095 -1.396168644 -1.356431821
#> 96  096 -0.026075207 -0.968987471
#> 97  097 -0.391720369  0.754262378
#> 98  098 -0.690218747  0.705461083
#> 99  099 -0.108683372 -0.112574101
#> 100 100  0.193562385 -1.546250133

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.19400504 -0.497568260
#> 2   002 -1.80928452 -0.318185654
#> 3   003  0.63740850 -0.464758722
#> 4   004 -1.55730052 -0.103612509
#> 5   005 -0.25396461 -0.308523515
#> 6   006 -0.34256978 -2.504633149
#> 7   007 -1.54918108 -0.235788852
#> 8   008  0.41593277 -1.012230611
#> 9   009 -0.54522267  0.472760806
#> 10  010 -0.88860712 -0.445075460
#> 11  011  2.40648908  1.802322660
#> 12  012  0.27044030 -0.015245914
#> 13  013  0.42398851  0.006800583
#> 14  014  0.56916275  1.283092912
#> 15  015  1.02159289  0.323004581
#> 16  016  0.04752258  1.579390599
#> 17  017  1.35549008  0.262788326
#> 18  018  0.58363216  2.080885599
#> 19  019 -0.27272582  1.180414272
#> 20  020 -0.90619080 -1.070769137
#> 21  021 -0.25550936 -0.341567968
#> 22  022 -0.19374198 -1.592234139
#> 23  023  0.60277816  0.707723041
#> 24  024 -1.10483898 -0.885934180
#> 25  025  0.54424445  0.533500647
#> 26  026 -1.21712601 -0.487417329
#> 27  027  0.05291497  0.269501745
#> 28  028 -0.28644922  0.138212136
#> 29  029  0.74577205  0.027445163
#> 30  030  0.08249809  0.186424964
#> 31  031  0.64832883  1.109519950
#> 32  032  0.83524629  1.920675903
#> 33  033 -0.21129232 -0.479297718
#> 34  034 -0.61103774 -0.070806551
#> 35  035  0.31874541 -1.141760422
#> 36  036 -1.03299405 -0.532863875
#> 37  037 -1.90980739 -1.576588590
#> 38  038 -0.01461137  0.009555995
#> 39  039 -0.66836159  0.421898884
#> 40  040 -0.30726937  0.651082121
#> 41  041  2.15809319  0.695650119
#> 42  042 -0.86604391 -0.190445947
#> 43  043  1.94822894  0.933534018
#> 44  044 -1.37592132 -1.749446696
#> 45  045  0.15056752  0.533914481
#> 46  046  1.02866531  0.559518656
#> 47  047  0.56205979 -0.738815470
#> 48  048 -0.88781464  1.060939933
#> 49  049  1.87754664  0.913307429
#> 50  050  1.03039325  1.354342298
#> 51  051  1.69735981  0.087460027
#> 52  052  0.88360341 -0.166602190
#> 53  053  1.16979363  0.441917658
#> 54  054 -0.74459708  0.223117028
#> 55  055  2.09755023  1.941178581
#> 56  056 -1.14381730 -0.950614780
#> 57  057 -1.62767943  0.634819809
#> 58  058  0.13837881 -1.493472454
#> 59  059  0.09428687 -0.271678286
#> 60  060 -0.09188073 -0.305786325
#> 61  061 -0.10440621  0.902869782
#> 62  062 -1.42151900  0.136780337
#> 63  063 -0.72096964 -0.757015863
#> 64  064  0.57102664  0.350284890
#> 65  065  1.40754150  0.889507120
#> 66  066  1.59086153  2.197438950
#> 67  067 -0.46187335  0.795369575
#> 68  068  0.25798327 -0.741835674
#> 69  069 -1.00477756  0.184070615
#> 70  070  1.11822035 -0.842612957
#> 71  071 -0.36850455 -0.312850157
#> 72  072 -0.67936566 -1.150638285
#> 73  073  0.02110568 -0.774423827
#> 74  074 -1.34737798  0.605078201
#> 75  075  0.22623670  0.502324021
#> 76  076 -0.28025332 -0.484170714
#> 77  077  0.06368992  0.393749005
#> 78  078 -0.47383549  0.730613511
#> 79  079 -1.49557804 -1.798771142
#> 80  080  1.04491752  0.283681460
#> 81  081  2.09990898  0.502176279
#> 82  082 -1.20849172 -0.089723002
#> 83  083  1.64773461  0.022413962
#> 84  084 -0.47482418 -0.764959570
#> 85  085 -1.89687996 -0.567545205
#> 86  086  0.64930441  0.368005291
#> 87  087 -1.18911865 -0.447365448
#> 88  088 -0.17392368  0.454273535
#> 89  089 -0.21445625  1.059267300
#> 90  090  0.69012583  1.592910630
#> 91  091 -2.55488723 -1.527437707
#> 92  092 -0.31568043 -1.715547509
#> 93  093  0.42251582 -1.576061026
#> 94  094 -0.98996642 -0.661581861
#> 95  095  1.15142232 -1.118297990
#> 96  096 -0.68181105  0.494808395
#> 97  097  0.34600026 -0.155994825
#> 98  098  0.92584707  0.456412066
#> 99  099 -0.61289537  0.091913470
#> 100 100  0.31091969 -1.499360902