Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4950406
#> Y_2 0.4950406 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2065418  0.5107182
#> 2 -1.0653628 -0.6132460
#> 3 -0.7852473 -0.4856378
#> 4  1.0673798 -0.2933268
#> 5  0.2127304  0.4988988

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.26305080  0.68304014
#> 2   002  0.93597316  0.68508838
#> 3   003  0.96077220  1.68252295
#> 4   004 -2.40933566 -0.13745855
#> 5   005 -0.45311514 -0.30604468
#> 6   006  0.03612231  0.56885962
#> 7   007  1.49555383 -0.50728226
#> 8   008  0.42914376 -0.23035910
#> 9   009  1.94929759  1.24372062
#> 10  010  1.18795968  1.86030689
#> 11  011  0.43107927 -0.28329130
#> 12  012 -0.73882068 -0.59623406
#> 13  013  1.58719017  0.24331187
#> 14  014  0.67963498  0.34441669
#> 15  015 -0.90487834  0.55879953
#> 16  016  0.53212281 -1.04905160
#> 17  017  0.61138832 -0.41924405
#> 18  018  1.90180411  0.85146786
#> 19  019  1.42521262  0.93962652
#> 20  020 -1.05182852 -0.90710244
#> 21  021 -0.26442420  0.14991492
#> 22  022  2.23647598  0.82594471
#> 23  023 -0.39694224 -1.12708207
#> 24  024  0.25309441  0.76956186
#> 25  025  0.18158780  0.73087494
#> 26  026  0.54566880  0.27539976
#> 27  027  0.88505190  1.12955152
#> 28  028 -0.08074110  0.16127607
#> 29  029 -0.63320930 -0.04613784
#> 30  030  0.48824416  1.82134799
#> 31  031  1.25083040  1.43242594
#> 32  032 -0.75827205 -2.09543761
#> 33  033 -0.13760096  0.29270098
#> 34  034  1.15613300  2.39421470
#> 35  035 -1.58924190 -0.64655628
#> 36  036  0.05893454  0.65673764
#> 37  037 -0.10103743 -0.87524392
#> 38  038  0.37922290  0.53415113
#> 39  039 -0.23799621 -1.05478907
#> 40  040 -1.02813227 -0.62354212
#> 41  041 -0.20171473  1.42927949
#> 42  042 -0.94266553  1.35491141
#> 43  043  0.15921109  0.07376009
#> 44  044  1.54238071 -0.01711754
#> 45  045  1.21901616  1.88628731
#> 46  046  1.65116043  2.41017268
#> 47  047 -0.03797233 -0.70853885
#> 48  048  0.44089594 -0.27368253
#> 49  049 -0.40813619  0.11380317
#> 50  050 -1.33815865 -0.59520151
#> 51  051 -0.33654096 -1.12230302
#> 52  052 -1.07311794 -0.09690965
#> 53  053 -0.81972532  0.39330205
#> 54  054  0.36607406 -0.25677143
#> 55  055 -0.54216943  0.35180061
#> 56  056 -2.60627862  0.06731666
#> 57  057 -0.40850570  0.84509016
#> 58  058  0.71818929  1.00741947
#> 59  059  0.05733090  0.49426355
#> 60  060  1.48288634  0.16431031
#> 61  061  0.06572160 -0.38154298
#> 62  062  0.61690101  0.90292251
#> 63  063  0.89612713  0.07932620
#> 64  064  0.64110379  0.52189438
#> 65  065 -1.27653777  0.58247448
#> 66  066 -2.27995171 -1.28042773
#> 67  067  0.76557887  0.91210961
#> 68  068  0.24745271  0.39446126
#> 69  069 -1.33623740 -1.36543248
#> 70  070  1.08628694  0.40627404
#> 71  071 -1.03383313 -1.44336353
#> 72  072  0.56034072 -2.08664846
#> 73  073  1.85024056  0.51581126
#> 74  074 -0.30357981  0.69574093
#> 75  075 -0.50496569 -0.22597808
#> 76  076  0.13531270 -1.56192246
#> 77  077  0.60952240 -0.91302581
#> 78  078  0.21143073 -0.26984205
#> 79  079  1.96325159  1.41087570
#> 80  080 -2.19977987 -1.77571497
#> 81  081 -0.23225059 -0.06704210
#> 82  082  1.23683469 -1.57004828
#> 83  083  0.08295142  0.21382456
#> 84  084  1.05460941  0.10295802
#> 85  085  1.56121348  0.22432711
#> 86  086 -0.56150975 -1.82244774
#> 87  087 -1.23925996 -0.20651735
#> 88  088 -0.22382557 -0.37772832
#> 89  089  0.71099944 -0.22186757
#> 90  090  0.92839177  1.33456440
#> 91  091 -0.18902332 -1.16343237
#> 92  092  0.72251612  0.87223239
#> 93  093 -0.17717356  0.85260717
#> 94  094  0.07036178 -1.28426436
#> 95  095  0.39143632  0.71731226
#> 96  096  0.66411683  0.14891431
#> 97  097 -0.64427782  0.21402186
#> 98  098 -1.46463521 -1.37630568
#> 99  099  2.71362134  0.39586625
#> 100 100  0.09159010 -1.14129241

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.31557924 -1.16590769
#> 2   002  1.17049984  0.70634906
#> 3   003 -0.84284831  0.28304884
#> 4   004  0.20925749 -0.28841888
#> 5   005  1.35132638  1.01237006
#> 6   006  1.12028559  1.09340957
#> 7   007 -0.82283272  0.13468242
#> 8   008  1.37370681 -0.25333980
#> 9   009  1.40304617 -0.74586519
#> 10  010 -1.38931201  0.03750605
#> 11  011 -2.07388950 -0.94230578
#> 12  012  0.18501775  0.35700474
#> 13  013  0.47191728  0.43937961
#> 14  014 -0.03547413  0.22823803
#> 15  015  0.89306578  0.40815147
#> 16  016  0.61280113  1.34673801
#> 17  017 -0.28151606 -1.02606691
#> 18  018  0.77731850  0.45103116
#> 19  019 -0.07493281 -0.20157198
#> 20  020  2.11280067  0.91539316
#> 21  021 -0.69263857 -1.95450712
#> 22  022 -0.92545673 -1.39151808
#> 23  023  1.34691252  0.27930740
#> 24  024 -0.94720209 -0.38592045
#> 25  025 -0.61491245  0.62468527
#> 26  026  0.18527162  1.22320027
#> 27  027  1.46524852  1.01912145
#> 28  028 -1.03731808 -1.80673610
#> 29  029 -0.10458492  0.55731174
#> 30  030 -1.02961436  0.45057817
#> 31  031 -1.21718247 -0.68587132
#> 32  032 -1.08997362  0.03233072
#> 33  033  0.47808924  0.67661055
#> 34  034 -0.16420673  0.06718168
#> 35  035  0.83435998  0.84485661
#> 36  036  1.28117585 -0.32726767
#> 37  037  0.20890565  0.13892708
#> 38  038  2.04583284  1.02010065
#> 39  039 -0.58641097 -0.88405585
#> 40  040  0.07983461  0.24856637
#> 41  041 -1.05758262 -0.19331303
#> 42  042  0.80520299 -1.49158381
#> 43  043  0.23590288  0.05428412
#> 44  044 -1.03107091 -0.96529547
#> 45  045 -1.41343104 -0.06491311
#> 46  046 -0.59664206 -0.18065755
#> 47  047  0.13460813  1.15394087
#> 48  048  0.65481689 -0.05904354
#> 49  049 -1.44746273 -1.21094463
#> 50  050 -1.50612504  0.53265393
#> 51  051 -1.21963576 -1.27258179
#> 52  052  1.80524019  0.09422343
#> 53  053 -0.73593630  1.03146541
#> 54  054  2.64290415  1.94214329
#> 55  055 -1.11640534 -0.97341857
#> 56  056  1.68939417  1.68139291
#> 57  057 -0.31112451  0.48596368
#> 58  058  1.05854132  0.17650752
#> 59  059  0.21114765  0.35796153
#> 60  060  1.89687254  0.17494550
#> 61  061  1.75382571  1.75274517
#> 62  062 -1.01914069  0.59333448
#> 63  063  0.48845206  0.80056291
#> 64  064  1.40013536  0.59275700
#> 65  065 -0.59678820  0.13286801
#> 66  066 -0.95319541 -2.43956515
#> 67  067  0.49268251  0.88396253
#> 68  068  0.27456065  0.10117854
#> 69  069  0.08467294  0.30797108
#> 70  070 -0.46633707  0.43266322
#> 71  071 -0.32920767 -1.12695005
#> 72  072 -0.76871155  0.20784990
#> 73  073 -0.86366388 -0.13901988
#> 74  074  2.34248084  1.85050480
#> 75  075  1.33578020  1.87607788
#> 76  076  0.34927074  1.12387109
#> 77  077  0.76912189 -0.59027891
#> 78  078 -2.66571690 -2.75577926
#> 79  079  0.45424377  0.24471740
#> 80  080 -0.78331038  0.94333533
#> 81  081  0.34116661  0.03680787
#> 82  082  0.64279640  0.16076268
#> 83  083  2.13193772  1.25677035
#> 84  084 -0.93994726 -1.56291652
#> 85  085 -0.97603948 -2.15321187
#> 86  086  1.30149091  1.45595354
#> 87  087  0.49558912 -0.75519595
#> 88  088  0.38597495  0.99426408
#> 89  089  0.98391217  1.29152038
#> 90  090  1.96599367  0.09504995
#> 91  091  1.49806155  0.76150424
#> 92  092  1.26092874 -0.06181505
#> 93  093  0.91739208 -0.55433539
#> 94  094 -1.04432784  1.60018253
#> 95  095  0.85913659  0.33741084
#> 96  096  1.30203867  0.49372101
#> 97  097  0.36282341  0.92935951
#> 98  098 -1.26408826 -1.28116209
#> 99  099  0.94690301  1.40706114
#> 100 100  1.33891724  1.89776483