Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5121972
#> Y_2 0.5121972 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.6546202 -0.6938076
#> 2  2.4480111  1.3382797
#> 3 -0.3932007 -2.1757483
#> 4  0.9137947  0.0726814
#> 5  0.2177285  0.5265339

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.99115505 -0.22953933
#> 2   002  2.22094552  0.97447844
#> 3   003 -0.46135514 -0.85699338
#> 4   004 -0.30468276 -1.94688705
#> 5   005  0.24261454 -0.43330944
#> 6   006 -1.13887929  0.37223637
#> 7   007  0.20812835 -0.68788328
#> 8   008  0.47204799 -0.43508677
#> 9   009  0.82584728  1.71523952
#> 10  010  1.10068023 -0.50007101
#> 11  011 -0.95577621 -0.06966172
#> 12  012  0.49738790  0.83719708
#> 13  013 -0.65834929  0.65632071
#> 14  014  1.94730082  1.18990093
#> 15  015  0.68811806  0.58881837
#> 16  016  0.79129767 -0.09588675
#> 17  017 -0.44883433 -0.49561879
#> 18  018 -0.41055294  1.36692470
#> 19  019 -0.02506982  1.36507460
#> 20  020 -0.35237816 -0.76697787
#> 21  021 -0.82982162  1.76842084
#> 22  022 -0.14084410  0.04486555
#> 23  023  1.22494592  2.15384499
#> 24  024  1.25042091  1.36114638
#> 25  025 -1.01005559  0.23278181
#> 26  026  0.50917455  0.30803683
#> 27  027  0.07549699 -0.42198861
#> 28  028 -2.93179534 -2.95699035
#> 29  029  0.50873973  1.01519385
#> 30  030 -0.75498617  0.38273073
#> 31  031 -1.38256078  1.00489322
#> 32  032  1.97815383  1.32748291
#> 33  033 -0.94886252 -0.93966454
#> 34  034 -0.31783708 -0.67569906
#> 35  035  0.15812713 -0.35143882
#> 36  036  0.95131509 -1.07876882
#> 37  037 -0.70734751 -2.06280652
#> 38  038 -0.08943891  0.58444029
#> 39  039  0.86171924  0.29164234
#> 40  040 -0.52527433  0.78914679
#> 41  041 -0.71555368  1.09081170
#> 42  042 -0.11333240  0.87906867
#> 43  043 -0.33250363 -1.79327383
#> 44  044  0.01743891  0.97963224
#> 45  045 -1.51445843 -2.30548036
#> 46  046  0.49020247 -0.58997443
#> 47  047 -1.50604918  0.91721404
#> 48  048  0.32879574  0.20017602
#> 49  049  1.12723992  1.10859746
#> 50  050 -0.37105039  1.21925164
#> 51  051 -0.60937774  0.63408780
#> 52  052 -0.62801249 -0.09524054
#> 53  053  0.41832547 -0.25267423
#> 54  054 -1.32457205 -2.21806020
#> 55  055 -0.54086686 -0.20511176
#> 56  056  0.42449353 -0.21814483
#> 57  057  0.11960168 -1.33668086
#> 58  058 -0.30564385 -1.24332822
#> 59  059  0.19905769 -0.63344846
#> 60  060 -1.26871001 -3.37758582
#> 61  061  0.67681431 -0.41741182
#> 62  062  1.35431449 -0.87631580
#> 63  063  2.31640069  2.20371834
#> 64  064  0.13567491  1.01670330
#> 65  065 -0.89344870 -1.53608299
#> 66  066  0.34065967  0.26054536
#> 67  067  1.05519389 -0.06805602
#> 68  068 -0.81050510  0.36206458
#> 69  069  0.35734274 -0.53852521
#> 70  070 -2.17351048 -1.34952993
#> 71  071  0.11957299  1.49803911
#> 72  072  1.59089799  1.30434140
#> 73  073 -0.37544941  1.23282436
#> 74  074  0.49885600  0.02303637
#> 75  075  0.40898905  0.95050503
#> 76  076 -1.19152330  0.31649228
#> 77  077  0.24349676  1.10573289
#> 78  078  0.12198339 -1.46936024
#> 79  079 -0.60511332 -0.07864138
#> 80  080  0.71544415  2.23221378
#> 81  081 -0.53656383  0.04517475
#> 82  082  0.30504915 -0.39088949
#> 83  083 -0.93205077 -0.59557864
#> 84  084 -0.16134665  0.37183693
#> 85  085 -1.14184050 -1.85436664
#> 86  086 -1.27360861 -0.44511019
#> 87  087  0.41559653  0.79051256
#> 88  088  1.18245225 -0.67058221
#> 89  089  0.49100421 -0.53110415
#> 90  090  0.07212745  0.65000478
#> 91  091 -0.22499640  0.21779914
#> 92  092  0.34495392  1.23869958
#> 93  093  0.28341831  0.16850622
#> 94  094  1.39213872  0.60360012
#> 95  095  0.99643363 -0.55012099
#> 96  096 -0.02919379 -0.99763964
#> 97  097 -0.16607030 -0.16856215
#> 98  098  0.07712286  1.35081744
#> 99  099  1.02692844 -0.20322093
#> 100 100  0.61415964 -0.43837571

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -0.659867056  0.676479495
#> 2   002  1.568710463  0.316806562
#> 3   003  0.090634519 -0.244461646
#> 4   004 -0.859000377  1.519012057
#> 5   005  0.505171547  0.998369438
#> 6   006  0.462763022  0.568107796
#> 7   007 -0.199165249 -0.874338508
#> 8   008  1.813769231  1.161139178
#> 9   009  1.014337637  0.085977392
#> 10  010 -1.687616519 -1.540226731
#> 11  011  0.026246706  0.796839740
#> 12  012 -1.512298813 -0.676704314
#> 13  013  1.117330631 -0.402815570
#> 14  014  0.637059616  0.450131925
#> 15  015 -1.374761584 -2.632058593
#> 16  016  0.037373732 -2.315586536
#> 17  017  0.151197241 -1.261198565
#> 18  018 -0.832618562  0.406324352
#> 19  019  0.940983791 -1.026424737
#> 20  020  0.855550068  0.340125417
#> 21  021 -0.683035652 -0.927853161
#> 22  022  0.342700433 -0.357640786
#> 23  023 -1.221921634  0.086869514
#> 24  024 -1.051990433 -1.154854797
#> 25  025  0.021437797  0.542369874
#> 26  026  0.305778908 -0.392812247
#> 27  027  2.264151436 -1.085187942
#> 28  028  0.885523312  0.861676532
#> 29  029 -1.482820099 -1.119197558
#> 30  030  1.097587937  0.733243463
#> 31  031  0.073432693  0.863142217
#> 32  032 -1.177297201  0.186839471
#> 33  033 -1.519101173 -0.550961525
#> 34  034 -0.400135100 -1.128978173
#> 35  035  0.611091767  0.342285096
#> 36  036 -0.061978203  0.048542706
#> 37  037  2.056974263  1.591854509
#> 38  038  0.777618171 -0.554499783
#> 39  039  1.281722876  1.765827273
#> 40  040 -0.552021274 -1.112241508
#> 41  041  0.127982637 -2.028455396
#> 42  042  0.884908686 -0.442929969
#> 43  043  1.509637914  0.300819071
#> 44  044 -0.145258477  0.011820723
#> 45  045 -0.823077186  0.755058691
#> 46  046 -0.527682740  0.056407496
#> 47  047  1.872731738  1.980699791
#> 48  048 -0.490456468 -0.926403342
#> 49  049 -0.761239049 -0.032137499
#> 50  050  0.001526992  0.364146233
#> 51  051 -1.667869266  0.123814240
#> 52  052 -0.056004416  0.354909830
#> 53  053  0.287329710 -0.123268649
#> 54  054 -0.380785404  0.552640464
#> 55  055 -1.093109977 -0.697129633
#> 56  056 -0.270222326 -0.354328459
#> 57  057  0.336936583  0.833374967
#> 58  058  0.457042463  2.283416318
#> 59  059 -0.405600767 -0.001662553
#> 60  060 -0.097239411 -1.597820578
#> 61  061 -0.612075789  0.025333100
#> 62  062 -0.145535081 -0.065460775
#> 63  063 -0.296696482  0.010698567
#> 64  064  0.024311877  1.201108632
#> 65  065  1.565587996  0.914517496
#> 66  066  0.164385052  0.004655899
#> 67  067  1.243577544  1.873732236
#> 68  068 -0.433313699 -0.757100539
#> 69  069  0.106948347  0.400673380
#> 70  070 -0.334334697  1.245343655
#> 71  071 -0.700241977 -0.730251944
#> 72  072  0.309028498 -0.286931856
#> 73  073 -0.259612911  0.616969070
#> 74  074  0.116625135  1.876000744
#> 75  075 -0.703267094 -0.531081666
#> 76  076 -0.325343209 -0.644766164
#> 77  077  0.607380863  0.235447518
#> 78  078  0.359490058 -0.719112490
#> 79  079 -0.868186248 -1.330358281
#> 80  080 -1.433277322 -1.907248076
#> 81  081  1.081990104 -0.161911888
#> 82  082  0.289108492 -0.897111603
#> 83  083  1.984156525 -0.111352474
#> 84  084  0.057794444  0.244723789
#> 85  085 -0.756541223 -1.439454062
#> 86  086 -0.762412182 -1.184293020
#> 87  087  0.764322235  0.987850890
#> 88  088 -0.722337001 -0.172944399
#> 89  089 -0.208520136 -0.157083774
#> 90  090 -0.244139820  1.094244057
#> 91  091  0.436707244  0.938027520
#> 92  092 -0.795913128 -0.882853555
#> 93  093 -0.239765781  1.807885657
#> 94  094  0.369207323  0.430765977
#> 95  095  0.323143210 -0.415952459
#> 96  096 -0.135871966  0.964353517
#> 97  097 -0.248658458  0.631495026
#> 98  098  0.147106698 -0.394409297
#> 99  099 -0.160676719 -0.342620478
#> 100 100 -1.859541617 -1.485075081