Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4916507
#> Y_2 0.4916507 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  1.35081593 -0.19993941
#> 2  0.68692596  0.03627892
#> 3 -1.17595585 -1.20450419
#> 4 -0.28829460 -1.11499131
#> 5  0.06801048  0.15890071

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.98845467 -0.95670507
#> 2   002 -0.50153486 -0.41569924
#> 3   003  1.53451952  1.69076488
#> 4   004 -1.57036288 -2.35342445
#> 5   005 -1.31880698  0.22796023
#> 6   006 -0.78869878  0.15039119
#> 7   007 -1.22270099 -1.37467951
#> 8   008 -0.04070824  0.03464652
#> 9   009  0.28198996  1.69812123
#> 10  010  0.04367353 -2.04414399
#> 11  011 -1.87612741 -0.29426989
#> 12  012  0.54706044 -1.79707188
#> 13  013 -0.77514948 -1.90923666
#> 14  014  0.68181779  0.36108483
#> 15  015  0.51362194  0.45437623
#> 16  016  1.02880135  0.36242233
#> 17  017  0.16572543 -0.79334057
#> 18  018 -0.75840913  0.83830525
#> 19  019  0.46202404 -0.82558439
#> 20  020 -0.50698683 -1.58199536
#> 21  021  0.06977128  0.25424763
#> 22  022  0.38647064 -0.88897404
#> 23  023 -0.97218748 -0.17789982
#> 24  024  1.63365207  0.60972477
#> 25  025 -0.15679574  0.83868978
#> 26  026 -1.25879039 -0.75768489
#> 27  027  1.46175783  0.29814905
#> 28  028  1.16772085  1.15651766
#> 29  029 -1.01500670  0.18582954
#> 30  030  0.28117044  0.50353300
#> 31  031 -2.05426972 -2.19384544
#> 32  032 -1.76444904 -1.02564380
#> 33  033 -0.08083047 -0.27417626
#> 34  034 -1.00468270 -0.68468103
#> 35  035  0.36690870  0.98028134
#> 36  036 -0.79390643 -0.31406391
#> 37  037 -0.59010206 -0.19895581
#> 38  038 -1.17475327  0.48404871
#> 39  039 -1.46642316 -0.36310144
#> 40  040  0.67936128  0.52865461
#> 41  041  3.12906709  2.14214007
#> 42  042 -0.40536652  0.77267321
#> 43  043 -0.69930242 -1.48869181
#> 44  044  0.02001607 -0.85190612
#> 45  045  0.35287152 -0.35030901
#> 46  046 -0.01737299 -0.69351314
#> 47  047 -2.24081743 -1.48395702
#> 48  048 -0.17292134 -0.91122999
#> 49  049  0.16538850 -0.11754201
#> 50  050 -0.50849031 -0.41007208
#> 51  051  1.26364449 -0.36271682
#> 52  052  0.74310457  1.30854547
#> 53  053  0.24300396  0.03009255
#> 54  054 -0.88084814 -2.08931511
#> 55  055  0.30434834 -0.44822496
#> 56  056  0.34385402 -0.29526732
#> 57  057  0.09209095  0.92564067
#> 58  058 -0.45897514 -1.87428413
#> 59  059  1.45068414 -0.04868339
#> 60  060 -0.72776014 -1.74221956
#> 61  061 -0.16219570 -0.64115654
#> 62  062 -1.04138954 -0.13554808
#> 63  063  0.83491965 -0.15301494
#> 64  064  0.04337427 -0.45113223
#> 65  065 -1.71808904 -0.91901992
#> 66  066  0.34994231 -0.23745100
#> 67  067  0.24926459 -1.57456127
#> 68  068 -1.47038420  0.08723060
#> 69  069  1.12141895  1.13242244
#> 70  070 -1.10637920  1.56311012
#> 71  071 -0.23904969 -0.81647177
#> 72  072 -0.30247118 -0.11361118
#> 73  073  1.00858732 -0.44555118
#> 74  074 -0.44382025 -0.43801113
#> 75  075 -0.23261583 -0.52879621
#> 76  076  1.68997973  1.23441788
#> 77  077  1.40999864  0.89030184
#> 78  078 -0.31991367  0.51225847
#> 79  079 -0.87609148 -0.09306523
#> 80  080  0.90717384 -0.67360457
#> 81  081  0.76834468  0.84773999
#> 82  082 -1.83826244  0.16838448
#> 83  083  2.42451637  0.59176202
#> 84  084  2.29456538  0.76709498
#> 85  085 -0.08440775  0.81577151
#> 86  086  0.06180833 -0.59133334
#> 87  087  0.73400377  1.82004629
#> 88  088  1.33931532 -0.44646986
#> 89  089  0.13350330  1.11478744
#> 90  090 -2.33518064 -2.58399372
#> 91  091  0.61821203 -1.14435700
#> 92  092 -0.32104263  0.32921442
#> 93  093  1.41067624  1.19002660
#> 94  094  1.17335850  0.95084767
#> 95  095 -0.60365468  0.16958605
#> 96  096  0.11697761 -1.89541564
#> 97  097 -0.47742886 -1.41875584
#> 98  098 -0.58015291 -0.58354870
#> 99  099  0.51897967  0.52365409
#> 100 100  0.19228712  0.17989860

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  1.55531230  1.829578751
#> 2   002 -0.56414929 -0.725677165
#> 3   003  0.09470354  0.062975709
#> 4   004  0.22703863  1.239181535
#> 5   005  1.69844777  2.143163234
#> 6   006 -0.95637718 -1.429429510
#> 7   007 -1.00990590 -0.570544924
#> 8   008  0.74069327  0.184905778
#> 9   009 -1.19849292 -1.943080745
#> 10  010  1.63354036 -0.110428509
#> 11  011  1.71385076  2.085869929
#> 12  012 -0.19405921  0.061556128
#> 13  013 -0.44808023  0.251679251
#> 14  014  1.41834323 -0.027465928
#> 15  015 -0.76756772 -0.653813063
#> 16  016 -1.17670244 -1.228253961
#> 17  017 -0.42355596  0.583317533
#> 18  018 -0.49788891 -1.339949378
#> 19  019 -0.82424049 -0.280456010
#> 20  020  0.95223032 -0.141400630
#> 21  021 -0.00712612 -0.572586875
#> 22  022  1.15575062  1.633538190
#> 23  023  0.20451974 -0.547407177
#> 24  024 -1.08457829 -0.775350536
#> 25  025 -0.57163084  0.960199364
#> 26  026  0.83336061  0.131587504
#> 27  027  0.25181401  0.639707590
#> 28  028 -0.02968090  0.071420921
#> 29  029 -0.02968838  0.350912567
#> 30  030 -0.44623389 -0.362284888
#> 31  031  1.33602506  1.411567052
#> 32  032  0.59277276 -0.285002725
#> 33  033 -0.27534704 -0.997099623
#> 34  034  2.21164996 -0.020518278
#> 35  035 -0.48999532 -0.253433374
#> 36  036  1.13640215 -0.141582542
#> 37  037 -0.16764328  0.962538202
#> 38  038 -0.73759201  0.293272219
#> 39  039 -0.50359139  0.671655194
#> 40  040  0.26133662 -0.513372041
#> 41  041 -1.50056663 -0.484681944
#> 42  042  0.64735419  0.767571194
#> 43  043 -0.23630313  0.018824826
#> 44  044 -0.80160193 -1.957563692
#> 45  045 -0.45619684 -0.006773242
#> 46  046 -0.78931256 -0.119775528
#> 47  047  0.35740105 -0.489852756
#> 48  048 -0.11946586  0.437569149
#> 49  049 -0.73418589 -0.912675710
#> 50  050  1.35488576  0.941847048
#> 51  051  0.12230932 -0.838023921
#> 52  052 -0.26917553 -0.430681295
#> 53  053 -0.87827660 -1.315511429
#> 54  054 -0.54629611  0.233483344
#> 55  055 -0.60454003  0.195440837
#> 56  056  0.84542499  0.853737527
#> 57  057 -1.30024808 -1.153178003
#> 58  058 -0.01840379 -2.025532787
#> 59  059 -1.64138450 -0.340464896
#> 60  060 -0.35613558  0.066383918
#> 61  061  1.55026528  1.825305308
#> 62  062 -0.28766396 -0.403259277
#> 63  063  2.36107975  1.211237026
#> 64  064  0.79233213  1.327455512
#> 65  065 -0.03214194 -0.901548196
#> 66  066  0.54907206  1.228156808
#> 67  067  1.21856340  2.044926366
#> 68  068 -0.25163782  0.378473339
#> 69  069  0.84891589  0.968303474
#> 70  070 -2.25382188 -2.092264061
#> 71  071 -0.11035742  0.044737424
#> 72  072  0.64483794  1.437544287
#> 73  073  1.42048629  0.630220569
#> 74  074 -0.03832762  1.323562531
#> 75  075  1.38618057 -0.293324563
#> 76  076 -0.40716070  0.208779606
#> 77  077  2.29504664  1.517478470
#> 78  078  1.53634845  1.066054847
#> 79  079 -1.69318523 -0.371533550
#> 80  080  0.18116663  0.878037986
#> 81  081  1.93969622  0.705849556
#> 82  082  0.59929973  0.139836584
#> 83  083 -0.75433753 -1.783883660
#> 84  084  1.18483872  2.299490151
#> 85  085 -0.27014065  0.921089479
#> 86  086 -1.55574593 -1.266418081
#> 87  087  0.86482393  0.033452261
#> 88  088  0.55458077  0.501194357
#> 89  089 -0.99111369 -1.903579558
#> 90  090 -0.18570886  1.698090972
#> 91  091  0.26792042 -0.384698622
#> 92  092 -1.00114401 -1.331672244
#> 93  093  0.22745630  0.805128463
#> 94  094  0.06816789 -0.696477574
#> 95  095 -1.00596764 -1.447878915
#> 96  096  0.76314426  0.196931975
#> 97  097  0.01476832  0.377553682
#> 98  098  1.10068033 -0.213093892
#> 99  099  1.01052114 -0.430902032
#> 100 100 -1.03715395 -0.723347019