Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4369223
#> Y_2 0.4369223 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -1.8571253 -1.8113550
#> 2 -0.3725999  0.2008889
#> 3 -0.2156709  0.3631743
#> 4  2.3914424  1.6793602
#> 5 -0.6705918 -1.0927240

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.195798262  0.884252442
#> 2   002  0.081810697  0.220192236
#> 3   003  2.822675816  0.690295385
#> 4   004  0.030212571  1.141489403
#> 5   005  0.027251798  0.375648970
#> 6   006  0.565201885  0.116199574
#> 7   007 -2.043140140 -1.063690499
#> 8   008 -1.393625107 -0.026070008
#> 9   009 -0.019768713  0.638103900
#> 10  010  0.238204386  1.039018669
#> 11  011 -0.317688782 -1.121302359
#> 12  012 -0.579490889 -1.229178611
#> 13  013 -1.259359392  0.932106004
#> 14  014  0.166737592 -0.519639089
#> 15  015 -0.508906531 -1.232622504
#> 16  016 -0.378351342  2.436389848
#> 17  017 -0.499349469 -1.415591936
#> 18  018  0.382496854  0.809352917
#> 19  019 -1.148982625 -2.028098176
#> 20  020 -2.258711543  0.254125704
#> 21  021 -0.855727528 -1.216243842
#> 22  022 -0.204764453 -0.507615391
#> 23  023 -0.137285323  0.322093845
#> 24  024  0.145177526  0.707613627
#> 25  025  0.252297791  2.037711644
#> 26  026 -0.567524852  0.807652654
#> 27  027 -3.274480692 -2.091007090
#> 28  028 -2.675162797 -1.757142864
#> 29  029  2.250474537  1.561870594
#> 30  030 -0.477517292 -1.840231608
#> 31  031 -0.112273726  0.081419658
#> 32  032 -1.841930417 -1.749763123
#> 33  033 -0.256630753  0.214683954
#> 34  034  0.179912683  0.600154489
#> 35  035  1.199557190 -0.341969054
#> 36  036  0.142895939 -0.623540669
#> 37  037  1.136322683 -0.301584881
#> 38  038 -0.936499381 -2.445878006
#> 39  039 -0.422927502 -0.643860817
#> 40  040 -0.297659092 -0.044985871
#> 41  041 -0.278702832  2.240028473
#> 42  042  0.127405401 -1.269910260
#> 43  043  0.362937625  1.349393323
#> 44  044  0.172259045 -0.488986995
#> 45  045 -1.746092479 -0.134403258
#> 46  046 -0.296152446 -0.043031638
#> 47  047  0.163878062 -0.408491992
#> 48  048 -1.382902453 -0.118438792
#> 49  049 -0.317758294 -0.618918489
#> 50  050  1.352186937  0.749876116
#> 51  051  0.478870114 -0.080875166
#> 52  052 -0.510028997 -1.551592571
#> 53  053  0.468031492  1.111282509
#> 54  054  0.423113466  0.084236133
#> 55  055  0.848577537  0.876722545
#> 56  056  0.311830387 -0.860619838
#> 57  057  1.311074162 -0.250285711
#> 58  058  0.222512702  2.223957254
#> 59  059 -0.042899314 -1.633986913
#> 60  060 -0.058227222  0.048727340
#> 61  061 -1.756252707 -1.578246839
#> 62  062 -0.783114885 -0.722029161
#> 63  063  0.366567375 -0.171172406
#> 64  064  1.993717294  1.377330856
#> 65  065 -0.477273558 -1.423982667
#> 66  066 -0.033844005  1.117273459
#> 67  067  0.698068507  2.256889545
#> 68  068 -0.004941245 -0.457567571
#> 69  069  0.338713483 -0.619932555
#> 70  070  1.189181054  0.284455885
#> 71  071 -1.011458601 -1.301618120
#> 72  072 -1.029737597  0.448935513
#> 73  073 -0.330224898 -0.090841548
#> 74  074  0.180256464 -0.302071692
#> 75  075 -0.799184932 -1.384009936
#> 76  076 -0.221938302 -0.008275882
#> 77  077 -1.116593533 -1.579058973
#> 78  078  0.433527911 -0.034055762
#> 79  079 -1.524933137  1.503227758
#> 80  080  0.933297403  0.049705368
#> 81  081  0.118964188 -0.795719397
#> 82  082  0.238252586 -0.026035576
#> 83  083 -0.335117269  0.936566287
#> 84  084 -1.789957412  0.506539400
#> 85  085  0.177449533  0.076895380
#> 86  086  0.859350603 -0.504849996
#> 87  087  0.858451926 -0.174647385
#> 88  088  1.128433209  1.689466959
#> 89  089 -1.791490720 -1.685781209
#> 90  090  1.294217833  0.356535187
#> 91  091 -1.503408473 -2.029242610
#> 92  092  2.229181210  0.192600488
#> 93  093  0.849571749  0.601428531
#> 94  094  1.260566876  0.500137277
#> 95  095  1.002281349  0.205182803
#> 96  096  0.293063493  1.342219797
#> 97  097  0.826953070  1.064715470
#> 98  098 -0.154813755  0.298483336
#> 99  099 -0.175667761 -0.356619930
#> 100 100  0.139070809 -1.197703932

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  1.184343e+00  0.59947791
#> 2   002 -8.467083e-01 -0.51229870
#> 3   003 -1.806295e+00 -0.78123164
#> 4   004 -1.337106e+00 -2.04670058
#> 5   005  1.395410e-01 -0.27397413
#> 6   006 -8.793157e-01 -0.10721230
#> 7   007  2.678520e+00  1.80155825
#> 8   008  2.493211e+00  2.24322788
#> 9   009  9.779966e-01  0.65035104
#> 10  010  1.683478e+00  1.59274386
#> 11  011 -2.612507e-01 -0.74841332
#> 12  012 -9.249342e-01  0.34135998
#> 13  013 -1.514195e-01 -0.90845593
#> 14  014 -1.228838e+00 -1.43274005
#> 15  015  5.240223e-01  2.01471377
#> 16  016  1.006522e+00  0.25726543
#> 17  017 -2.852465e-01 -0.47499709
#> 18  018  5.543699e-01 -0.49604484
#> 19  019 -4.317158e-01 -1.79836832
#> 20  020 -1.330088e+00 -0.89619918
#> 21  021 -6.796682e-01 -0.53364670
#> 22  022 -1.247211e+00 -1.34200799
#> 23  023 -1.240565e+00 -0.51894478
#> 24  024 -6.973344e-01 -0.94059971
#> 25  025  2.133764e-01  0.21043620
#> 26  026 -1.677661e-01  1.08098271
#> 27  027  1.470290e+00 -0.34484813
#> 28  028  1.359342e+00  0.50747419
#> 29  029 -5.814375e-01 -1.33794858
#> 30  030 -6.111694e-01 -1.34014312
#> 31  031 -1.115739e+00 -1.73865924
#> 32  032 -1.644124e+00 -0.51611079
#> 33  033 -2.597758e-01 -1.07229444
#> 34  034  1.239464e+00  0.49156361
#> 35  035 -4.265861e-01 -0.17966632
#> 36  036  4.623771e-01  0.25124080
#> 37  037  2.671972e+00 -0.23255778
#> 38  038  1.345620e+00  0.47237334
#> 39  039 -1.402263e+00 -2.25007965
#> 40  040 -2.131161e+00 -0.63637537
#> 41  041 -5.457202e-01 -2.51301307
#> 42  042  1.543292e-01  0.48680627
#> 43  043 -6.903034e-05 -1.40898895
#> 44  044 -1.180082e-01  2.22981551
#> 45  045  1.452306e-01 -1.08703371
#> 46  046 -4.936261e-01  1.38133050
#> 47  047 -2.942456e-01 -0.20037610
#> 48  048  5.637563e-01  1.38076702
#> 49  049 -3.153593e-01  0.82421017
#> 50  050 -5.677559e-01  0.11425308
#> 51  051 -5.897933e-01 -1.35531347
#> 52  052 -1.088606e+00  0.26368496
#> 53  053 -7.983979e-01 -0.56144720
#> 54  054 -1.255371e+00 -0.54162792
#> 55  055 -7.503008e-01  0.60517283
#> 56  056 -9.990205e-02  0.53741711
#> 57  057  1.846087e+00  1.40091998
#> 58  058  9.681507e-03  0.11528495
#> 59  059 -7.501957e-01  0.29516674
#> 60  060  7.901059e-01  0.28370157
#> 61  061 -1.715179e+00 -2.07708545
#> 62  062  2.129105e-01  0.84479581
#> 63  063  1.881086e-01  0.07328394
#> 64  064 -3.310740e-04  0.34704971
#> 65  065  1.251319e+00  1.06286781
#> 66  066  7.047108e-01 -0.36612347
#> 67  067 -1.256227e+00 -0.69871311
#> 68  068 -6.835430e-01 -0.62642114
#> 69  069 -1.900780e-01 -1.55129915
#> 70  070  7.526541e-01  0.53150714
#> 71  071 -6.243801e-01 -0.32748954
#> 72  072 -1.431875e+00 -0.39640236
#> 73  073 -7.534498e-01 -0.12776446
#> 74  074  3.150330e-01  1.18006451
#> 75  075  1.607167e-01 -0.87612450
#> 76  076  7.820022e-01  0.59912081
#> 77  077 -7.650910e-01  0.34466608
#> 78  078 -1.814371e+00 -0.57016638
#> 79  079  2.008797e-01  0.07075609
#> 80  080 -2.053299e-01  0.35638803
#> 81  081  1.028193e+00  0.24993681
#> 82  082 -1.760383e-01 -0.26202289
#> 83  083 -1.181535e-01  0.42989676
#> 84  084  1.187962e+00 -1.34206411
#> 85  085  1.463784e+00  0.10566997
#> 86  086 -2.041688e+00 -2.38824259
#> 87  087 -1.044192e-01 -0.36829546
#> 88  088  4.113116e-01 -0.64899180
#> 89  089 -8.454904e-01 -1.57677087
#> 90  090  8.294338e-01  0.44776234
#> 91  091 -7.827895e-01  0.71355479
#> 92  092  3.558277e-01  0.02430795
#> 93  093 -6.185744e-01 -0.67559603
#> 94  094  1.859529e-02  0.07754008
#> 95  095 -4.555583e-01 -0.53162969
#> 96  096 -1.287225e+00 -0.26828300
#> 97  097 -6.545375e-01 -0.00780103
#> 98  098 -1.153676e+00  0.15540284
#> 99  099  1.152554e+00  0.42543495
#> 100 100  1.141500e+00 -0.22941775