Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4944411
#> Y_2 0.4944411 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  0.50728340 -0.03321775
#> 2 -0.22790575 -0.40967378
#> 3 -0.07853469 -0.93314747
#> 4 -0.30056808 -0.58076941
#> 5  0.39354446  0.42647516

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.704332186 -0.37585734
#> 2   002 -1.507990794 -2.05816815
#> 3   003  0.122135676  0.79313097
#> 4   004 -0.291309336  0.30976619
#> 5   005  0.694597895 -0.03735633
#> 6   006  0.635648927  0.38538338
#> 7   007  2.082728730  0.52767379
#> 8   008 -0.753594232 -1.08157523
#> 9   009 -0.803727136 -0.99684479
#> 10  010 -1.659563924 -2.25555074
#> 11  011  1.639243325  0.25303705
#> 12  012 -0.186170664 -0.23964312
#> 13  013 -0.978119066 -1.82456940
#> 14  014 -0.814691858 -1.32345745
#> 15  015 -1.015303534  0.15253111
#> 16  016  1.289597073  0.91339718
#> 17  017  1.553613353  1.04967989
#> 18  018  0.247959386 -0.13794843
#> 19  019 -0.289050791  0.33670058
#> 20  020 -0.184097200  0.20778450
#> 21  021 -0.981559263 -0.32425699
#> 22  022  0.875829196  1.15804333
#> 23  023 -0.699068929 -0.18359030
#> 24  024  0.426833508 -1.53877243
#> 25  025 -1.271688913  1.76036588
#> 26  026 -1.991880958 -2.01553708
#> 27  027  0.348491964  0.08778620
#> 28  028  0.503264727 -0.73238515
#> 29  029  0.884001960  0.50458653
#> 30  030 -0.931197618 -0.37287399
#> 31  031  1.406437520  0.68425947
#> 32  032 -0.314513034  0.85812554
#> 33  033 -0.129526762 -0.17124542
#> 34  034  1.077631780  2.18086702
#> 35  035  2.455918119  1.64561930
#> 36  036  0.795208409 -0.60843577
#> 37  037  0.212337150  0.65081526
#> 38  038  0.047169320 -1.88264862
#> 39  039  1.121825579  0.90157761
#> 40  040 -0.105510449  0.78361248
#> 41  041  0.249212202  0.92960738
#> 42  042  0.963063741 -0.46921860
#> 43  043  0.305081135  0.41228803
#> 44  044 -1.660780588 -2.36967990
#> 45  045  0.326350268  0.95475953
#> 46  046  0.184632056  0.75765223
#> 47  047 -1.289954414 -1.29142597
#> 48  048 -0.292397288 -1.51793224
#> 49  049  1.270423814  0.64398951
#> 50  050 -0.405043650  0.84093690
#> 51  051  0.862769275 -0.28237280
#> 52  052 -0.024085132  1.80863489
#> 53  053  1.432349423  0.21518693
#> 54  054 -0.379759931  1.69994014
#> 55  055 -0.732770429  0.05502076
#> 56  056 -0.048033775 -0.63208681
#> 57  057  0.120358519 -0.91066431
#> 58  058 -0.662704011  0.10047064
#> 59  059  0.774770705  0.30696773
#> 60  060  0.121316423 -1.53272542
#> 61  061  0.474275868 -0.41996899
#> 62  062 -0.797379576 -0.90390004
#> 63  063 -0.204221726 -0.53779016
#> 64  064 -0.103115704  0.44608223
#> 65  065  1.409265276  1.05379422
#> 66  066 -0.791707400 -1.74419862
#> 67  067  0.290408041 -0.83424971
#> 68  068  0.024410096  0.40343558
#> 69  069  0.601340782 -0.15795079
#> 70  070  1.749246870  1.37750071
#> 71  071  0.318713310  0.10225721
#> 72  072  0.230160538 -1.40960027
#> 73  073 -0.458724956  1.49064154
#> 74  074  0.035720521  2.35008520
#> 75  075  0.289793337  0.36179373
#> 76  076  0.860894037  0.71672877
#> 77  077 -0.905597281 -0.00818319
#> 78  078 -0.106377213 -0.92425449
#> 79  079 -0.712869412 -1.14015998
#> 80  080  1.165525010 -0.66253275
#> 81  081  0.683332389  0.85530045
#> 82  082 -0.448714762 -0.17688000
#> 83  083 -0.154225156  0.78387805
#> 84  084 -1.592883662 -1.19325737
#> 85  085 -0.619762094 -0.92053931
#> 86  086 -0.166636455  0.17581142
#> 87  087 -1.250611000 -0.05239185
#> 88  088  0.004456375 -0.51018181
#> 89  089  0.269879000  1.12907675
#> 90  090 -2.096133304  0.15714573
#> 91  091  0.883901624 -0.10875230
#> 92  092  0.101088005 -0.76973899
#> 93  093  0.131721398  1.12529815
#> 94  094  0.168318213  0.10882236
#> 95  095  0.667246413  0.93488422
#> 96  096  0.027296325  0.00521831
#> 97  097  0.237977880 -1.32154799
#> 98  098  0.834399862  0.97746772
#> 99  099 -0.990193381 -0.69743755
#> 100 100  0.705653439 -0.74751487

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.145988421  1.14619816
#> 2   002  0.191088980  1.43966369
#> 3   003 -0.876842576 -0.26263057
#> 4   004 -0.443450512 -0.64987139
#> 5   005  1.461715883  0.67331838
#> 6   006  0.382724771  0.30076010
#> 7   007  1.705626944  1.20566351
#> 8   008  1.293986064 -0.26256119
#> 9   009 -1.126110464 -0.66514927
#> 10  010  1.038144409  1.46370544
#> 11  011 -1.187337141  0.65140897
#> 12  012  1.003075620  0.29572706
#> 13  013 -1.057050626 -1.47006918
#> 14  014  0.202355317  1.81602637
#> 15  015 -1.274939698 -0.55798374
#> 16  016 -0.887548388  0.20124333
#> 17  017 -1.563902019 -0.64744811
#> 18  018 -1.895203442 -1.22605948
#> 19  019  0.879021487  2.46447851
#> 20  020 -0.233294708  0.60145476
#> 21  021  0.139291308 -0.31160150
#> 22  022  0.653432119 -0.96041059
#> 23  023  1.841595814  0.78516080
#> 24  024 -0.067354923  0.19890067
#> 25  025  0.076550217 -0.79332704
#> 26  026 -1.526974084 -0.03438391
#> 27  027 -0.978064139  0.17888236
#> 28  028  0.374823500 -1.04770523
#> 29  029 -0.397534143 -1.40965416
#> 30  030 -1.048020034 -1.24796164
#> 31  031  0.232559044 -1.55261203
#> 32  032  0.232084970  0.68426273
#> 33  033  1.594296137  1.34147217
#> 34  034  2.209461270  0.88238029
#> 35  035  0.357148608 -0.83075786
#> 36  036  0.088368207  0.42160664
#> 37  037  1.265915519  0.25033552
#> 38  038  0.620199592  0.95352490
#> 39  039 -0.712130521 -0.97950987
#> 40  040  1.007184319  1.43589969
#> 41  041 -0.752972426 -1.57643738
#> 42  042 -2.087688681 -2.27414258
#> 43  043 -0.412186362  0.13113215
#> 44  044  0.417234313  0.26807959
#> 45  045  0.171244796  1.44954125
#> 46  046 -0.982879739 -0.28012776
#> 47  047 -0.902333422  0.45703816
#> 48  048  0.558465869  0.66623040
#> 49  049  0.055608652 -1.09803825
#> 50  050  0.148668528  0.38512865
#> 51  051  0.245077552 -0.71993989
#> 52  052 -0.062130048 -0.45157557
#> 53  053 -0.774822135 -1.80471392
#> 54  054 -0.057452578  0.25718914
#> 55  055  1.142711516 -0.15600027
#> 56  056 -1.496747381 -1.00340049
#> 57  057  0.831545069  0.25885176
#> 58  058  0.582866724  1.58301542
#> 59  059  0.398951865  1.04778986
#> 60  060 -0.255819096  0.94382826
#> 61  061  0.813752373 -1.33214470
#> 62  062 -1.303727591  0.78094982
#> 63  063 -0.595142970  0.60752544
#> 64  064 -0.813566170 -0.84140551
#> 65  065 -0.009336786 -0.62654135
#> 66  066  2.321804123  0.25250837
#> 67  067 -2.546622963 -2.00942613
#> 68  068 -0.627471422 -0.32686013
#> 69  069  0.011999852  0.60241899
#> 70  070 -0.303539670  0.49736531
#> 71  071 -1.280434390 -1.40358773
#> 72  072 -0.775093213 -1.31535631
#> 73  073 -1.671730881 -0.67961041
#> 74  074 -0.864495665 -0.54491853
#> 75  075 -1.986503646 -0.93073375
#> 76  076 -1.352050625 -0.80369978
#> 77  077 -0.953932460  1.11429832
#> 78  078 -0.545618232  0.15775413
#> 79  079  0.509476851  0.42467773
#> 80  080  0.951841705  1.41676443
#> 81  081  0.573114574 -0.09913877
#> 82  082 -1.637976361 -0.57975762
#> 83  083 -0.446915108  0.15983017
#> 84  084 -0.813357618 -2.01122251
#> 85  085  0.221087480  0.17535874
#> 86  086 -0.300843129  0.92183549
#> 87  087  0.456188627  1.91973397
#> 88  088  0.879150747  0.56604544
#> 89  089  0.093226523  0.84352412
#> 90  090 -0.460035527  0.18216025
#> 91  091 -0.341721025 -0.80984353
#> 92  092  1.490530061  1.03302237
#> 93  093  0.082554437 -0.31660540
#> 94  094 -0.623220342 -0.91552577
#> 95  095  0.597847999 -0.91899283
#> 96  096 -0.587617497 -1.87543472
#> 97  097  0.379033798  1.30275807
#> 98  098 -1.277213013 -0.08879095
#> 99  099  0.716335962  0.31695046
#> 100 100  1.772781214  2.51036839