Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5267948
#> Y_2 0.5267948 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.8240854 -0.1272078
#> 2  1.1101540  0.7469179
#> 3  1.4694455 -1.6296472
#> 4  1.3734829  1.7488721
#> 5  1.7363768  1.7271779

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1           Y_2
#> 1   001 -1.069395901 -1.2420187762
#> 2   002 -1.281790434 -0.1908186118
#> 3   003 -0.850825283  0.1572285690
#> 4   004 -0.781273846 -0.5968001813
#> 5   005 -1.599737343  0.4536797706
#> 6   006 -1.776726374  0.0234999197
#> 7   007 -0.736013123  1.5199255968
#> 8   008 -1.108064584  0.0153904630
#> 9   009 -0.258047655  0.0835535991
#> 10  010  1.718354688  1.6602862503
#> 11  011  0.040349684 -0.1461301976
#> 12  012  0.280551130 -0.1700994491
#> 13  013  1.472316179 -0.2028592288
#> 14  014  1.991694106  0.4500300824
#> 15  015  0.359550095  1.0850804404
#> 16  016  0.644869883  0.2967663227
#> 17  017  0.565768329 -0.0843275057
#> 18  018 -0.979047165 -1.8137866960
#> 19  019 -0.833748830  0.4372541238
#> 20  020 -0.911343707 -0.4755400875
#> 21  021  2.022046308 -0.1395785197
#> 22  022  0.353377719 -0.2869984561
#> 23  023  1.394969158 -0.1676703184
#> 24  024 -0.117422317  0.0687853930
#> 25  025  2.017253201  1.1476652692
#> 26  026  0.900780874  0.0007111695
#> 27  027  0.172338468  0.2264456618
#> 28  028 -1.094111532 -0.3580075736
#> 29  029 -1.594077843  0.0005107882
#> 30  030  1.146186248  1.8435524151
#> 31  031 -1.852892933 -1.4875168600
#> 32  032  1.933230593  1.5096612037
#> 33  033 -0.662698705 -1.4158048294
#> 34  034 -0.935543089 -0.3636999429
#> 35  035  0.582682931 -0.9602389631
#> 36  036 -0.122228282  1.4248284676
#> 37  037  0.320002973 -1.7488152492
#> 38  038 -1.447683957 -0.0927537109
#> 39  039  0.604805632  1.7850003308
#> 40  040  0.178047575  0.2041661482
#> 41  041 -0.187322856  1.0144460302
#> 42  042 -0.397061525  1.1525078397
#> 43  043  0.785811055  0.9687596604
#> 44  044  1.463241769  0.7986897841
#> 45  045  0.706679469 -0.1299365325
#> 46  046 -0.135589854  0.9166369032
#> 47  047 -0.860860867 -0.8908853939
#> 48  048 -0.371398045  0.3027291763
#> 49  049 -0.920648705 -0.4553857876
#> 50  050  1.512584597  0.9169406616
#> 51  051 -1.501269371 -0.4673933918
#> 52  052 -0.108182699  2.1761558410
#> 53  053 -0.009924275 -1.3265076286
#> 54  054 -1.987884498 -2.2039255276
#> 55  055 -0.900518310 -0.1611427510
#> 56  056  0.595211490 -1.4025659035
#> 57  057 -2.018005032 -3.1093436440
#> 58  058 -0.013005894 -1.2851507552
#> 59  059 -1.497268799 -0.7240431661
#> 60  060 -0.232750608 -0.7385835083
#> 61  061  1.111375410  1.5083244122
#> 62  062 -0.564141453  0.6582519236
#> 63  063  2.263779702  1.9527447393
#> 64  064  0.572472604  1.3942497528
#> 65  065  0.705653603  1.8034829469
#> 66  066  0.181048374  0.0731628819
#> 67  067  0.678844358  2.1686223468
#> 68  068  0.037358108 -0.2893709609
#> 69  069 -3.156457037 -2.0034547860
#> 70  070 -0.171549604  0.6662888268
#> 71  071  0.409969955  0.1590116180
#> 72  072 -1.702096452 -1.1948163257
#> 73  073 -0.157727515 -0.7267087499
#> 74  074 -0.416218728 -0.8803311259
#> 75  075  1.203624029  0.9425212753
#> 76  076  0.241381111  2.2793344060
#> 77  077  0.168900356  1.8361373407
#> 78  078  1.559552543  0.4176862864
#> 79  079  0.250804821 -0.5870459939
#> 80  080 -0.438128105  0.7144565987
#> 81  081 -1.802145156 -0.9501035936
#> 82  082 -0.961542117  0.4674657737
#> 83  083 -0.786589562 -0.8618461625
#> 84  084 -0.503384731 -1.7081869430
#> 85  085  1.443515028  0.5900222884
#> 86  086 -1.224273567 -2.0693056480
#> 87  087 -0.146919493  1.5964579626
#> 88  088  1.028416899  0.5894660582
#> 89  089 -1.694160271 -1.6692280048
#> 90  090 -0.049295588 -1.2461382494
#> 91  091  1.522625265  0.7545635443
#> 92  092 -0.509189355 -1.5536982074
#> 93  093 -0.623450471 -0.3700604133
#> 94  094  0.064603494  1.4582527605
#> 95  095  0.958353817  0.7351013924
#> 96  096 -1.173469315 -1.0118205086
#> 97  097 -0.670161932 -1.2521272953
#> 98  098  0.325643663  2.0414554860
#> 99  099  0.538351516 -0.5833751854
#> 100 100  0.107288331  0.8308744590

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.363871753 -0.13377476
#> 2   002 -0.435942344 -0.05582144
#> 3   003 -0.417232491 -1.17616109
#> 4   004  0.926030543  2.47710455
#> 5   005 -0.006187059  1.66988227
#> 6   006  0.265117126  0.73194217
#> 7   007 -1.002627754 -0.89402378
#> 8   008  0.391910703  0.62509616
#> 9   009  0.129118402  0.81732333
#> 10  010  1.161701383  1.23367769
#> 11  011 -0.079415127 -0.26193220
#> 12  012 -1.183621433 -0.76940551
#> 13  013  1.973289925  0.13642399
#> 14  014 -1.073804150 -1.27220876
#> 15  015  1.146137049 -0.76164381
#> 16  016 -1.383899198 -1.13714994
#> 17  017  0.603354232 -0.03407669
#> 18  018 -0.498499503 -1.30338917
#> 19  019 -0.420437535  0.12729557
#> 20  020 -1.721893224 -0.37596161
#> 21  021 -0.536775790  0.83103058
#> 22  022 -0.028493165  0.11551642
#> 23  023  0.309185383  0.95669393
#> 24  024  0.345053887 -1.33398630
#> 25  025 -1.654991063 -0.43778374
#> 26  026 -0.756756315 -0.82299735
#> 27  027 -0.735089717  0.01919336
#> 28  028 -1.400377392  0.13759020
#> 29  029  0.838853741 -0.09176668
#> 30  030 -0.098411964 -0.83648233
#> 31  031 -0.147291926  0.31241539
#> 32  032  0.674309680 -1.57004874
#> 33  033 -0.867843120  0.21870991
#> 34  034  0.601958244 -2.51674592
#> 35  035  0.222638029  0.08572032
#> 36  036  1.039180195  0.72994769
#> 37  037 -2.411583801 -1.02737969
#> 38  038 -0.273343986 -0.12110721
#> 39  039  0.614087766  0.73987789
#> 40  040 -0.944792517 -0.68505290
#> 41  041 -0.343851219 -0.74727336
#> 42  042 -1.626485898 -0.74520930
#> 43  043  1.035773107 -0.07373393
#> 44  044 -0.476055265 -1.16551618
#> 45  045 -1.157683664 -0.50991404
#> 46  046  1.222134053 -0.08231514
#> 47  047 -0.012104773  0.11752279
#> 48  048 -1.858571568 -1.24265000
#> 49  049 -0.596731909 -0.40249290
#> 50  050  0.725628485  1.33329089
#> 51  051 -2.386474386 -1.33265789
#> 52  052  0.437772634  1.83555425
#> 53  053 -1.162665026 -1.14493134
#> 54  054 -1.119995508 -0.70734551
#> 55  055 -0.336585884 -1.39862359
#> 56  056  0.061093639 -0.12882534
#> 57  057  1.367073068  0.44717012
#> 58  058  0.231558572  0.37774567
#> 59  059  0.522878789 -0.35714632
#> 60  060 -1.170473607 -1.71415164
#> 61  061 -1.328315053  0.15860016
#> 62  062 -1.680881835 -0.82379011
#> 63  063  0.608191541  0.46265082
#> 64  064  0.246446135 -0.78404919
#> 65  065  1.312950875  1.21086282
#> 66  066  2.229219307  1.61976744
#> 67  067 -0.075757315 -0.64231083
#> 68  068 -0.903095204  0.51665307
#> 69  069 -0.883980122  0.04340208
#> 70  070  0.175113120  0.08626237
#> 71  071  1.545701246  0.37301603
#> 72  072 -0.684181854 -1.01711934
#> 73  073  0.610150667 -0.56669042
#> 74  074  0.711540668  0.31014178
#> 75  075 -0.745620887 -0.78928625
#> 76  076 -0.314973703  0.12210984
#> 77  077  0.023601919 -0.06982097
#> 78  078 -1.842536193 -1.01176086
#> 79  079  1.630134764  1.10657122
#> 80  080 -0.005884062 -0.77244840
#> 81  081 -0.533177648 -1.14382074
#> 82  082  1.414605517  0.77093163
#> 83  083  0.465143517  0.18978529
#> 84  084 -0.321544117 -0.11966520
#> 85  085  0.707689549 -0.20148516
#> 86  086 -0.669362960 -0.79059274
#> 87  087  1.420668162  0.61873986
#> 88  088 -0.793015994 -0.33789703
#> 89  089 -0.169259909  0.91544852
#> 90  090 -0.820430258 -0.48879844
#> 91  091  0.398710501  0.97774098
#> 92  092  1.202794550 -1.10483984
#> 93  093  0.633087381  0.68211863
#> 94  094  1.287120010  0.83118613
#> 95  095 -1.025437504 -1.46266961
#> 96  096  0.515404619  1.57238491
#> 97  097 -1.034109565 -1.19477049
#> 98  098 -0.815415724  0.70757901
#> 99  099 -2.720922797 -1.81158355
#> 100 100 -0.396273055 -0.75680089