Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4961247
#> Y_2 0.4961247 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.0786774 -2.35664108
#> 2 -0.8574770  0.66856344
#> 3  0.8356964  0.09394015
#> 4 -1.5493689  0.15552454
#> 5 -0.4710614  0.02836975

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.314397553  0.73443830
#> 2   002  0.588669369  1.06742302
#> 3   003  0.580150900  0.37033002
#> 4   004  1.950884402  0.21714648
#> 5   005 -0.847066929  0.33765442
#> 6   006  1.283800857 -0.96632689
#> 7   007  0.846157440  1.44978501
#> 8   008 -0.599884222 -0.08022771
#> 9   009  0.120970014  0.09246008
#> 10  010 -0.066147865  0.09871686
#> 11  011  0.255415514  0.58513395
#> 12  012  0.629387236 -0.22248291
#> 13  013 -0.063265202 -0.15834033
#> 14  014 -0.730914460 -2.35213296
#> 15  015 -0.410283598 -1.05334028
#> 16  016 -0.041019637  0.70813036
#> 17  017 -0.436190387 -1.08700994
#> 18  018  0.283977166 -0.12883143
#> 19  019  0.989840533 -0.58580549
#> 20  020  0.053790778  1.67942434
#> 21  021  0.777746795  0.32063394
#> 22  022 -0.945227325 -0.86663807
#> 23  023 -0.958788259  0.10206514
#> 24  024  0.040520508  0.33061993
#> 25  025 -1.257389812  0.39216277
#> 26  026 -0.734212384  0.32361509
#> 27  027  0.759789948 -0.76087348
#> 28  028  1.237327003  0.66742019
#> 29  029 -1.327103651 -0.21205399
#> 30  030  0.005858439 -1.37727943
#> 31  031 -0.991777386 -0.19056116
#> 32  032  0.101321695  0.76830215
#> 33  033 -1.539183705 -0.53727829
#> 34  034 -1.738916653 -1.51733393
#> 35  035 -1.046339817 -0.02712373
#> 36  036  1.280023607 -0.03299143
#> 37  037 -1.250181161 -0.11054992
#> 38  038 -0.093014419 -0.32850078
#> 39  039 -1.445663034 -0.09549295
#> 40  040  0.310555586  1.05058645
#> 41  041  0.298030995 -0.96351683
#> 42  042  0.696741534  0.27299135
#> 43  043 -1.221518909 -0.18809889
#> 44  044 -0.773052263 -3.23617478
#> 45  045 -0.333394689 -0.27123823
#> 46  046 -0.550474535 -0.78305899
#> 47  047 -0.245118383 -1.81897741
#> 48  048 -0.864206036  0.15350896
#> 49  049 -1.113788169 -1.67937751
#> 50  050 -0.503478597  0.40048469
#> 51  051  2.300530020  1.25179908
#> 52  052  0.845847675  1.69672984
#> 53  053  1.514274275 -0.14381179
#> 54  054  0.368176458  0.17950497
#> 55  055  0.570963291 -0.23484741
#> 56  056  0.754519036  0.47684300
#> 57  057  1.870701085  0.69339588
#> 58  058  1.223970375 -0.01455524
#> 59  059  2.174284966  0.38698590
#> 60  060 -0.396828406  0.25786685
#> 61  061 -1.817100773 -0.31092445
#> 62  062 -0.121646653  0.40375485
#> 63  063 -1.614012279 -2.04363323
#> 64  064  1.680042617  3.46005930
#> 65  065 -0.979024826 -0.06827362
#> 66  066 -0.138300780 -1.29688363
#> 67  067  0.547680985  0.58826038
#> 68  068  0.931404521 -0.09064438
#> 69  069  0.270637201  1.05703656
#> 70  070 -0.255428582 -0.79078793
#> 71  071 -0.319178625  0.25848754
#> 72  072  1.207826472  0.11095395
#> 73  073 -0.583008469 -0.94952013
#> 74  074  0.061345160 -1.07970211
#> 75  075  0.616475628  1.00341307
#> 76  076 -0.375313793 -0.08034084
#> 77  077  0.064340036  0.28931734
#> 78  078 -0.007851885 -0.51258423
#> 79  079 -0.917832648  0.34009591
#> 80  080 -2.230664429 -0.18309625
#> 81  081  0.186881476  0.74380481
#> 82  082  0.858666364  0.05634522
#> 83  083  0.575117894  1.67268413
#> 84  084 -1.329722091 -0.01061909
#> 85  085 -0.068248770  1.09040265
#> 86  086 -0.500111241  0.41096555
#> 87  087  0.289590310  1.68821021
#> 88  088  0.403000633  0.62266413
#> 89  089 -0.194065234  0.44683040
#> 90  090  2.622091249  1.44186668
#> 91  091  1.181380888  2.04344970
#> 92  092  1.307276999  0.99278028
#> 93  093 -0.038190768  0.36485540
#> 94  094 -1.234136233  0.67318377
#> 95  095 -0.191200793  0.72431070
#> 96  096  1.733522051  0.65923712
#> 97  097 -0.024754755  2.08784696
#> 98  098  0.565867498  0.56795071
#> 99  099  0.671518793 -0.33307229
#> 100 100  1.259355758  1.80070030

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.266365619 -0.53885491
#> 2   002  0.969441117  0.65499226
#> 3   003 -1.927211182 -0.44399402
#> 4   004  1.700928523  1.33963141
#> 5   005 -0.116002081 -0.01963787
#> 6   006  0.996390324  0.02036680
#> 7   007  1.348972210  2.06465582
#> 8   008 -1.557962273 -1.88322483
#> 9   009  2.386891900  0.77346114
#> 10  010  0.190049566  0.57797280
#> 11  011  0.608046625 -0.26412014
#> 12  012 -1.593501200  0.83402376
#> 13  013 -0.529480009 -0.72809382
#> 14  014  0.015764376 -0.56713740
#> 15  015 -0.568596577  0.21543693
#> 16  016 -0.877568411 -1.51878509
#> 17  017 -1.139968978  0.29433924
#> 18  018 -0.600440357  0.27146524
#> 19  019 -0.164690883 -0.17003452
#> 20  020 -0.298454127  0.01911038
#> 21  021 -1.439577239 -1.07347406
#> 22  022  1.168409454  1.50852093
#> 23  023 -0.808162021 -0.27822287
#> 24  024 -0.378650251  0.38864545
#> 25  025 -0.687042752  0.14296138
#> 26  026 -0.348119313  0.43413470
#> 27  027  1.957048777  0.52592601
#> 28  028  2.546345753  0.66041713
#> 29  029  0.398567533 -0.24276234
#> 30  030 -0.272805974  0.53241463
#> 31  031  1.103261917  1.74161660
#> 32  032  1.262377227  0.23271149
#> 33  033 -0.047665609 -0.28735742
#> 34  034 -0.090637643  1.58204297
#> 35  035  0.396663084  0.81009630
#> 36  036 -0.146118522 -0.06091723
#> 37  037 -1.864769688  0.98956606
#> 38  038 -1.578032724 -1.68203764
#> 39  039 -0.744489028  0.15239671
#> 40  040  0.839633657  0.20585507
#> 41  041  2.018717046  0.70398855
#> 42  042  0.681437236  1.22118029
#> 43  043  1.001516341  0.28652573
#> 44  044 -0.303890166 -0.54169836
#> 45  045  0.457492512 -1.01395053
#> 46  046  0.959051736 -0.98155149
#> 47  047 -2.283364414 -1.42483120
#> 48  048  0.871654758  1.07675739
#> 49  049  0.915641377  1.61801970
#> 50  050  1.150995892  1.00903350
#> 51  051 -0.260032876 -0.49300801
#> 52  052 -0.636935334  0.30770391
#> 53  053 -0.716192020  0.06260248
#> 54  054 -0.126229011  0.37406381
#> 55  055  1.711944668  1.18405775
#> 56  056  0.736610360  0.12296017
#> 57  057  0.463569283 -0.59811689
#> 58  058 -0.538423322 -0.07074763
#> 59  059 -0.087153107 -0.13245264
#> 60  060  0.372187084 -1.29174932
#> 61  061  0.705592513  1.89496646
#> 62  062  1.110345760 -0.16142292
#> 63  063  0.295678198  1.17436884
#> 64  064 -0.020129938 -0.64086386
#> 65  065 -0.786442147 -0.66928719
#> 66  066 -0.352836607  0.66576802
#> 67  067  0.215165621 -1.17593477
#> 68  068 -0.910898748 -0.47608556
#> 69  069  1.214708191  0.27125321
#> 70  070 -0.450901986 -0.31320636
#> 71  071 -1.033198976 -0.02191879
#> 72  072 -0.189219547 -0.98604792
#> 73  073  0.565003306  0.43581319
#> 74  074  1.298903003  1.50547002
#> 75  075 -1.204244054 -1.97376591
#> 76  076  0.242418109  2.21389982
#> 77  077 -0.007307653  0.23132159
#> 78  078  0.348341517 -0.54887141
#> 79  079 -1.230055575 -0.55380813
#> 80  080  0.139850354  0.56295967
#> 81  081 -3.108516092 -2.34710230
#> 82  082 -0.831476384  0.57609767
#> 83  083 -1.692919188 -1.74781351
#> 84  084 -0.406348262  0.20179932
#> 85  085  0.089222736 -0.65912899
#> 86  086  0.828621656  0.03600891
#> 87  087  0.800945285  0.66240077
#> 88  088 -0.082146351  0.63036659
#> 89  089 -0.660081555 -0.12663178
#> 90  090 -0.490714425  1.43568790
#> 91  091  2.448129655  0.88124623
#> 92  092 -0.007089083  0.15511982
#> 93  093  1.003558218 -0.07022204
#> 94  094 -0.547059764 -0.55717680
#> 95  095  1.288326387  0.06706010
#> 96  096 -0.169647340  1.34851097
#> 97  097  1.015955229  1.05342746
#> 98  098 -0.606931868 -0.71775045
#> 99  099  1.286835098  1.50623035
#> 100 100  1.412645396  0.70970232