Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5518516
#> Y_2 0.5518516 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3490276  0.3225059
#> 2  0.8890906  1.3618776
#> 3  0.6351166  1.4192956
#> 4 -0.4300109 -0.1052043
#> 5 -0.8231795 -0.3486060

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.13529685  0.622019969
#> 2   002 -0.82387907 -0.451199432
#> 3   003  0.67273659 -0.190510406
#> 4   004 -1.45025849 -1.255643076
#> 5   005  1.71716312  0.748309549
#> 6   006  0.40703375 -0.035325631
#> 7   007  0.39654303 -1.347133596
#> 8   008 -0.19309341  0.246122399
#> 9   009  0.03465828  1.161695535
#> 10  010  0.48411367  0.879282984
#> 11  011 -0.43038896  0.553451529
#> 12  012  0.96410947  1.242431906
#> 13  013 -1.68379025 -2.643192292
#> 14  014 -1.82057748 -0.068285707
#> 15  015 -1.11921236  0.001057995
#> 16  016 -1.45540392 -0.559956823
#> 17  017 -0.95510416 -2.074677514
#> 18  018 -0.44774973  0.713516460
#> 19  019 -0.82791391  0.160003857
#> 20  020  0.29929078 -1.246903915
#> 21  021 -0.95159716 -1.360653377
#> 22  022 -1.11267052 -0.112692632
#> 23  023  1.82131475  0.318580243
#> 24  024 -0.68149843 -0.380142813
#> 25  025  0.43137536  1.376399623
#> 26  026 -0.87394958 -0.881604107
#> 27  027  0.44635165  0.129646629
#> 28  028  0.21666411  0.386951351
#> 29  029  0.10613668  1.680566598
#> 30  030 -0.43714333  0.071862304
#> 31  031  0.07860452  0.555186129
#> 32  032 -2.76285569 -0.476080437
#> 33  033  1.76936084  0.929401962
#> 34  034  1.55435855 -0.805027843
#> 35  035 -0.98478378  0.087464791
#> 36  036  1.03682059  0.476013765
#> 37  037  0.59187035 -0.489455109
#> 38  038 -0.49158231 -0.285303542
#> 39  039 -0.12709594  0.509058119
#> 40  040 -2.12551033 -0.830554902
#> 41  041 -0.72965975 -0.527785480
#> 42  042 -0.17647134  0.537850947
#> 43  043  0.13142876 -0.563796411
#> 44  044 -1.09203595 -0.517930709
#> 45  045  0.43716490  1.466431676
#> 46  046  0.32809011  1.251974252
#> 47  047  0.07658641 -0.492941776
#> 48  048 -0.85480691 -1.132648373
#> 49  049  0.77770794 -0.285840771
#> 50  050 -2.11977026 -0.366625753
#> 51  051  0.87260288 -0.632244133
#> 52  052 -1.14823853 -0.369116028
#> 53  053  0.68117916  0.612689373
#> 54  054  1.18657863  1.325845145
#> 55  055  2.09061674  1.690801752
#> 56  056 -1.23337735 -0.707119334
#> 57  057 -0.44440802  1.363421519
#> 58  058  0.61305396  1.005728915
#> 59  059 -0.96801808  1.041767183
#> 60  060  1.21653166  1.172738522
#> 61  061  1.26562256  1.109985402
#> 62  062 -0.05272138 -0.463595165
#> 63  063 -1.10544108 -1.672050254
#> 64  064 -0.53463607 -2.117751242
#> 65  065 -0.28102581 -0.023240001
#> 66  066 -0.60139900 -0.294277242
#> 67  067 -0.18899062  0.487066235
#> 68  068 -0.20263830 -1.139855146
#> 69  069 -1.39940446 -1.121743312
#> 70  070 -0.06348149 -0.203421662
#> 71  071 -0.28694022  0.175655510
#> 72  072 -0.16461208 -0.348992232
#> 73  073  0.53174684 -0.102662825
#> 74  074  0.92937115  1.893520031
#> 75  075 -0.16316996  0.359353425
#> 76  076 -0.50744223 -0.157227021
#> 77  077  1.38616357  0.879475962
#> 78  078 -0.74343119  0.820970594
#> 79  079 -0.33403894  0.347738000
#> 80  080 -0.74847034  0.306545731
#> 81  081 -0.14077613 -0.754573351
#> 82  082 -3.56017782 -2.352722173
#> 83  083 -0.01989990  0.237009627
#> 84  084 -1.42133082 -0.707235238
#> 85  085  0.33162594 -0.099049569
#> 86  086 -1.32254695 -0.094915627
#> 87  087  0.32033133  1.791467230
#> 88  088  0.57130387  0.184588721
#> 89  089  0.55305314 -1.231405230
#> 90  090 -0.77194427  0.344502947
#> 91  091 -0.18152330  0.490957498
#> 92  092  0.76932205 -0.456411645
#> 93  093  0.86638764  0.205645368
#> 94  094 -0.78087877  0.832214052
#> 95  095 -0.70109552 -0.230212146
#> 96  096 -1.33100813 -2.003303362
#> 97  097 -0.64630248 -0.125896053
#> 98  098 -0.75041031 -0.921800763
#> 99  099  0.05951199 -0.072151864
#> 100 100  1.41989196  0.482380304

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001 -0.052283808  2.139048044
#> 2   002 -0.844714917 -1.221354340
#> 3   003  0.941351067 -1.821286147
#> 4   004 -0.842186209 -0.488405770
#> 5   005  0.705575470 -0.801207931
#> 6   006 -0.261947929 -0.089679113
#> 7   007  0.817274632  0.461721054
#> 8   008  1.105890368  1.443486467
#> 9   009 -0.647125927 -1.868378870
#> 10  010 -0.382294800 -0.809914734
#> 11  011 -0.007849304  0.672725878
#> 12  012 -0.721362179 -0.368015075
#> 13  013 -0.437275164 -0.038383557
#> 14  014 -1.619899912 -1.479992983
#> 15  015 -1.427085114 -0.524973308
#> 16  016 -0.725785741 -2.145803156
#> 17  017  0.831801368 -0.481011713
#> 18  018 -0.499757547  0.786767557
#> 19  019 -0.459201081 -1.479029583
#> 20  020  1.358785375  0.705291683
#> 21  021  0.347472148 -1.059577530
#> 22  022  0.360501475 -0.834655692
#> 23  023  0.159246571 -0.530461522
#> 24  024 -2.411960685 -1.195849622
#> 25  025 -1.561542393 -0.476993782
#> 26  026  0.820287236  1.851405962
#> 27  027 -0.223221204  0.244234961
#> 28  028 -0.388207183  0.002610824
#> 29  029 -0.118853954 -0.337193827
#> 30  030  0.879809329  0.932281886
#> 31  031 -1.078028379 -2.372413055
#> 32  032 -0.626630992 -1.638696945
#> 33  033 -1.184816997 -0.046161208
#> 34  034 -0.117199381  0.463701629
#> 35  035  1.271447527  3.112931830
#> 36  036  0.616867696 -1.122930681
#> 37  037 -1.136874683 -0.216445386
#> 38  038 -0.628298956 -1.746533098
#> 39  039 -0.793737184 -0.237482171
#> 40  040 -1.589286542 -1.428184889
#> 41  041 -1.254538688 -1.111876211
#> 42  042 -0.482452269 -0.735312740
#> 43  043 -0.822399830 -0.770271085
#> 44  044  1.134810369  1.013054380
#> 45  045  1.202860762  0.569736800
#> 46  046  0.637114680 -0.201270737
#> 47  047  0.108660893  0.374270990
#> 48  048  1.743798213 -0.459885584
#> 49  049  2.227022455  0.845055906
#> 50  050  0.787182479  1.557106989
#> 51  051 -0.585455876 -0.422545124
#> 52  052 -1.310099794 -0.590350000
#> 53  053 -0.578929686 -1.536387865
#> 54  054  0.213741235  0.600029070
#> 55  055 -0.010865388  0.214741710
#> 56  056 -0.354359450 -0.394589261
#> 57  057  1.571353308  2.134521635
#> 58  058  0.429767586  0.822809655
#> 59  059  1.587609190  0.530199951
#> 60  060  1.391608452 -0.127370975
#> 61  061  1.312415590  1.990013068
#> 62  062 -2.273559735 -2.191137243
#> 63  063  2.810970309  2.284335333
#> 64  064 -1.224827639 -1.637490223
#> 65  065  1.086641522  1.008014195
#> 66  066  0.038631230  0.283063162
#> 67  067 -0.623313006 -0.429252632
#> 68  068 -0.407456604  0.576790015
#> 69  069 -3.152770174 -1.054047053
#> 70  070  0.742291963  1.200898540
#> 71  071 -1.996903936 -1.615726823
#> 72  072  0.451039293 -0.252714497
#> 73  073 -1.441377380 -0.614601981
#> 74  074 -0.282678735  0.683295760
#> 75  075 -0.331343352 -0.317266358
#> 76  076  0.630112308  1.198834692
#> 77  077  1.888041257  0.171715584
#> 78  078  0.566938328  0.765223719
#> 79  079  0.545687557  0.056280767
#> 80  080  0.909057034  0.080173121
#> 81  081  2.103705508  0.824484719
#> 82  082 -1.030083120 -0.665984684
#> 83  083 -1.305081566 -0.309789173
#> 84  084 -1.393465386  0.224919032
#> 85  085  1.810384053  0.343987525
#> 86  086  0.501252841 -0.082352662
#> 87  087 -0.981626826 -0.734612638
#> 88  088 -0.816713138  0.113008643
#> 89  089  0.485260092  1.290533947
#> 90  090  1.497599844  2.131326913
#> 91  091  0.189603674 -0.739830733
#> 92  092 -0.422346248 -0.983980814
#> 93  093  0.411734950 -0.810445041
#> 94  094 -0.336178345  0.899764306
#> 95  095 -0.077808508  0.854611490
#> 96  096  1.135005805 -0.049999227
#> 97  097  1.895030227  2.670815265
#> 98  098  2.098366944  1.228313162
#> 99  099 -1.529321117 -0.713103379
#> 100 100  0.142936189  0.616331705