Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4745969
#> Y_2 0.4745969 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -2.7599649 -1.46789483
#> 2 -0.3967229 -0.79517006
#> 3  0.4605448 -0.09873185
#> 4 -0.9537650  0.53387580
#> 5  1.0474730  2.26310508

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  1.194609788  2.05327505
#> 2   002  0.592657673 -0.94627873
#> 3   003 -0.228927113  0.58954475
#> 4   004  2.585804507  2.23672736
#> 5   005 -0.724849564 -1.14595591
#> 6   006  0.123036017  1.25282415
#> 7   007 -0.010498141 -1.15221434
#> 8   008  0.502280974  0.36463417
#> 9   009  0.414492778 -0.42695412
#> 10  010 -1.983517317 -3.48832077
#> 11  011  1.005162473  1.08997103
#> 12  012  1.236599840 -0.43878037
#> 13  013 -0.671278491  0.18372985
#> 14  014 -0.242231801 -0.92131105
#> 15  015 -0.770609650 -0.57856618
#> 16  016  0.241811861  0.52153147
#> 17  017  0.868487466  0.66528329
#> 18  018 -0.854380678 -0.99514737
#> 19  019  1.412733019  0.11369544
#> 20  020  0.341685069  0.39986313
#> 21  021  0.332045072  0.62527953
#> 22  022 -1.230560243  0.54015439
#> 23  023  0.212421395  1.24483191
#> 24  024 -0.923959219  0.03709185
#> 25  025 -0.571016201  0.69023581
#> 26  026 -0.210026374  0.94130858
#> 27  027  0.414878153  0.60040512
#> 28  028 -1.336858842  0.48145308
#> 29  029 -0.295764544  0.86179027
#> 30  030 -1.506304228 -0.17945646
#> 31  031  0.606222692  1.43223553
#> 32  032  1.068059726 -0.42226633
#> 33  033 -0.512002862  0.50503696
#> 34  034 -1.778900112 -0.96704799
#> 35  035  0.284378942  0.92930358
#> 36  036  0.506578537  1.46185043
#> 37  037  1.767038313 -0.38846846
#> 38  038  0.991656876  0.18976033
#> 39  039  1.020990764 -1.09454522
#> 40  040  0.469095173 -0.26574133
#> 41  041 -1.179514465 -0.77022458
#> 42  042 -0.975517104 -0.46975710
#> 43  043  0.752443277 -0.07327235
#> 44  044  0.835865491 -0.06050793
#> 45  045 -2.291678724  0.40361980
#> 46  046  0.705681568  0.26471738
#> 47  047  0.669444487  0.11949458
#> 48  048  0.736081168 -0.24815274
#> 49  049  0.450392306 -1.43180959
#> 50  050 -3.491550512 -2.52181896
#> 51  051 -1.164941616  0.14096730
#> 52  052 -1.043027814 -0.98865048
#> 53  053  0.796295913  0.40845981
#> 54  054  1.258046119  1.56501829
#> 55  055  0.483871551  0.22671083
#> 56  056  0.077297535 -0.36733906
#> 57  057  1.366153674  0.05321347
#> 58  058 -0.853334314 -0.38270265
#> 59  059 -0.867302615 -0.52390743
#> 60  060 -1.175306935 -1.55200458
#> 61  061  0.801558682 -0.69070001
#> 62  062 -1.192819081 -0.64256432
#> 63  063  1.395558364  1.58051591
#> 64  064 -0.763101618 -2.24686142
#> 65  065 -1.234157270 -1.05494469
#> 66  066  0.079396209  0.36897194
#> 67  067 -0.270153995 -0.42553991
#> 68  068  2.366282640  0.70910408
#> 69  069 -0.610750141  0.88871136
#> 70  070 -0.564120018  0.04207191
#> 71  071  0.023149082 -1.35217163
#> 72  072  1.438806341  2.50350881
#> 73  073  1.428054909  0.11298991
#> 74  074  1.356174419  0.18087563
#> 75  075 -0.516123655  0.25479232
#> 76  076 -0.162232642 -0.19786366
#> 77  077 -1.951856566 -1.93271702
#> 78  078  0.590221658  1.20551480
#> 79  079  1.454142108 -0.04365200
#> 80  080  0.383059352 -0.47244473
#> 81  081 -0.809216582 -0.48947529
#> 82  082  0.317846384 -2.20613333
#> 83  083  2.715277326  0.47131703
#> 84  084 -1.375200534 -1.61357113
#> 85  085 -0.575947382 -0.67436148
#> 86  086 -0.981184012 -1.27722231
#> 87  087  0.703869625  1.66509384
#> 88  088  0.984796339 -0.34766383
#> 89  089  0.880599472  0.87549914
#> 90  090 -0.574697731 -0.69334543
#> 91  091  0.997884512 -0.38713464
#> 92  092  1.172493388  1.47626633
#> 93  093 -1.371180042 -1.34864084
#> 94  094  0.268188936  0.19953698
#> 95  095 -0.549498467  1.14911862
#> 96  096  0.218799155  0.29497264
#> 97  097 -2.014678274 -0.63265423
#> 98  098 -0.006998556  0.10652025
#> 99  099  0.273539158  0.18277993
#> 100 100  1.045159218  0.83769202

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.288395189 -0.55172352
#> 2   002 -3.023686856 -1.61736270
#> 3   003  0.793329566  1.14890344
#> 4   004  1.420985526  0.67781221
#> 5   005  0.085770720  0.32400951
#> 6   006  1.825110724  0.42086658
#> 7   007  0.680952117  0.66221908
#> 8   008 -1.305649877 -0.76034685
#> 9   009  0.974646869 -1.76879179
#> 10  010 -0.954030108 -0.28541661
#> 11  011  1.236156920  1.11757999
#> 12  012  0.670357490  0.78786671
#> 13  013  1.789406645  1.20857720
#> 14  014  0.901743917 -0.78840170
#> 15  015  0.305946169  0.73973902
#> 16  016 -0.365199071 -0.79383853
#> 17  017 -0.481397845  0.50227385
#> 18  018 -0.888256630  0.16397591
#> 19  019  0.227404644 -0.37860290
#> 20  020  0.346001814 -1.08500278
#> 21  021 -1.155759974 -1.19454350
#> 22  022 -0.492092826 -0.60467998
#> 23  023  0.499126945 -0.42371896
#> 24  024 -0.211282902  0.73006994
#> 25  025 -0.605181001 -0.14337332
#> 26  026  0.965250233  0.33749084
#> 27  027 -1.218188791  0.31116889
#> 28  028  1.299093749 -0.64573270
#> 29  029 -0.710138386 -0.75081999
#> 30  030 -0.410510668  0.00186563
#> 31  031 -0.629479550 -0.05624288
#> 32  032  0.409731462  1.10914680
#> 33  033  1.307180398  1.27376017
#> 34  034 -0.055867165 -1.06784439
#> 35  035  0.643933677 -0.53140316
#> 36  036  1.148854005  1.29595344
#> 37  037  0.854070289  0.07019038
#> 38  038  1.753939384  0.97498321
#> 39  039 -0.714068487 -0.93147214
#> 40  040 -1.263739907 -2.30506405
#> 41  041  1.996595268  0.55459126
#> 42  042  0.167851965  0.96701327
#> 43  043  1.932745070  1.02883243
#> 44  044 -0.608128157  0.17074814
#> 45  045  1.432470853  0.17177429
#> 46  046  1.299356006  1.74908184
#> 47  047  1.286062684  0.26193642
#> 48  048  1.769622489  1.13633458
#> 49  049 -1.722317552  0.68052973
#> 50  050  0.390071609 -0.06901283
#> 51  051 -0.273271030 -0.53898000
#> 52  052  0.251376424 -0.94294388
#> 53  053 -0.296226055  0.12984085
#> 54  054 -0.068191770  0.67344618
#> 55  055  0.150695151  0.85853149
#> 56  056 -0.585979366  0.91040153
#> 57  057 -1.269246798 -1.31690753
#> 58  058  0.193840101 -0.79741503
#> 59  059  1.036673142 -0.58051045
#> 60  060  0.672967729  0.59951455
#> 61  061  2.047907824  1.63413722
#> 62  062  0.511106957  1.62105707
#> 63  063 -0.627389628  0.64106420
#> 64  064 -0.321814371  1.46648176
#> 65  065  0.335554496  0.89160702
#> 66  066 -0.544092070  0.61337833
#> 67  067 -0.509540289 -0.98033351
#> 68  068 -0.689831713 -0.20790901
#> 69  069  0.547292213  0.93814266
#> 70  070 -0.766837670 -0.95675865
#> 71  071 -0.747546462 -0.38536738
#> 72  072 -1.023333363 -0.58499222
#> 73  073  0.885501314  0.29940992
#> 74  074  0.810125616  1.84292995
#> 75  075 -0.392929689 -0.01855877
#> 76  076 -0.235097894  0.21842422
#> 77  077 -1.372815614  0.51660285
#> 78  078 -0.644119156 -0.41197825
#> 79  079 -0.929029155  1.54839984
#> 80  080 -0.682672104 -0.17626372
#> 81  081  0.009839883 -0.42085062
#> 82  082  0.872464415 -0.45370019
#> 83  083 -1.184378296 -0.13128885
#> 84  084  2.222717244  0.01293964
#> 85  085  0.995147086  0.30663091
#> 86  086 -0.561409116 -1.39684722
#> 87  087  0.012661608  0.62297032
#> 88  088  0.712953655  1.15490959
#> 89  089  0.497542356  0.97493414
#> 90  090 -1.679741568 -0.45180336
#> 91  091 -1.138617985 -0.77179335
#> 92  092  1.531212747  0.89388133
#> 93  093 -2.542041872 -0.93013264
#> 94  094  0.714449305 -0.14968177
#> 95  095  0.514483277  0.10888832
#> 96  096 -1.241571410  0.04823613
#> 97  097  1.290690292  0.50929813
#> 98  098 -0.929812948 -0.87325671
#> 99  099 -1.367826520 -1.01612186
#> 100 100 -0.079477645  0.30297918