Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5182509
#> Y_2 0.5182509 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1 -0.167857210 -2.3441144
#> 2  0.681921618  0.1426030
#> 3  0.402442787  0.4176658
#> 4 -0.003608979 -0.8171199
#> 5 -0.353425582 -0.2410978

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.048979258 -0.955879867
#> 2   002  0.477412359  0.135520510
#> 3   003  1.404649155  1.160216469
#> 4   004  0.350990372  1.217762632
#> 5   005  0.983031077  0.637494499
#> 6   006  1.474968640  1.090395362
#> 7   007 -0.496172219  0.409287551
#> 8   008  0.261571165  0.076736479
#> 9   009 -1.930428247 -1.156644814
#> 10  010 -1.091443108  0.439035231
#> 11  011 -0.051565939 -1.061563792
#> 12  012  0.045399825 -0.695528376
#> 13  013 -0.681273199 -1.746733448
#> 14  014  0.856209061 -0.325644856
#> 15  015  1.535323346  0.097796561
#> 16  016 -0.710901741  0.708936069
#> 17  017 -1.905422405 -0.220370498
#> 18  018  0.833662556  2.503102394
#> 19  019  0.547158964  0.169003550
#> 20  020  0.848782913 -0.057640696
#> 21  021  1.943533036  0.333931833
#> 22  022 -1.772590590 -0.524121216
#> 23  023 -2.682173939 -1.975897904
#> 24  024 -0.082066128  0.131477612
#> 25  025 -1.698129848  0.211783824
#> 26  026  1.384450326 -0.399641468
#> 27  027  0.422793382  0.487338881
#> 28  028 -0.090797499 -0.223189521
#> 29  029  1.095714165 -0.335295493
#> 30  030  1.387774664 -0.027602923
#> 31  031  1.387832417  1.077429778
#> 32  032 -0.162659430  0.720344965
#> 33  033 -1.216695372 -0.500054568
#> 34  034 -1.097856035 -0.623659621
#> 35  035 -0.411160339 -0.008631467
#> 36  036 -0.055948730  0.051007703
#> 37  037 -1.821250151 -1.517821678
#> 38  038 -0.903214245  0.636254298
#> 39  039  0.072443448  0.446446997
#> 40  040 -0.959076861 -0.121431408
#> 41  041 -2.207897158 -2.051815233
#> 42  042  0.847301858  1.765138216
#> 43  043 -0.589536369 -0.826861063
#> 44  044 -2.235992140 -1.906047963
#> 45  045 -0.484285384  0.454979148
#> 46  046 -0.572100258 -0.654548325
#> 47  047 -0.478899345  0.136933800
#> 48  048  0.195165002  0.377954231
#> 49  049  0.922047930  0.536705724
#> 50  050  0.416724893  0.809592959
#> 51  051  2.154619441  1.819265361
#> 52  052 -0.162762523 -0.170376262
#> 53  053 -0.184014356 -0.235895180
#> 54  054  0.057563049  0.799362016
#> 55  055 -0.848382611 -0.095697000
#> 56  056 -0.592150912 -1.887007592
#> 57  057 -2.758744065 -1.239111882
#> 58  058 -1.175724410 -1.150462041
#> 59  059  0.560177029  1.949499194
#> 60  060  1.570613117  2.048001433
#> 61  061 -0.456385530 -0.160405995
#> 62  062 -0.406837018  0.116486643
#> 63  063  1.815254891  1.549428908
#> 64  064 -1.314491986 -0.640906152
#> 65  065  0.692765245  2.500073457
#> 66  066  1.551252331 -1.189898389
#> 67  067 -1.695166265 -1.178705064
#> 68  068 -0.654847705  1.264418312
#> 69  069  0.747776653  0.114008054
#> 70  070 -0.113171682  0.702140195
#> 71  071  0.502771745  1.031948632
#> 72  072 -0.004390328 -0.167794695
#> 73  073  0.030784475 -0.139605632
#> 74  074 -1.673155235 -1.688158885
#> 75  075  0.546342646  0.323565242
#> 76  076 -0.193232325  0.304393294
#> 77  077 -1.036173581 -0.768106045
#> 78  078  1.074784550  2.364529281
#> 79  079 -0.020970837  0.086450812
#> 80  080  0.062391585 -0.390993873
#> 81  081 -0.477527809 -0.897443202
#> 82  082  0.187375511  0.078623766
#> 83  083  0.801161897 -0.183970365
#> 84  084  1.050463085  0.968403552
#> 85  085 -0.800639125 -1.924594428
#> 86  086  0.792684912 -0.026836329
#> 87  087  2.380195355  0.733873911
#> 88  088 -1.110327654 -2.499338732
#> 89  089  0.431851335 -1.290650553
#> 90  090 -0.985343623 -0.874551662
#> 91  091 -0.124563920 -2.559138497
#> 92  092  0.124171395  1.459500928
#> 93  093  0.041767182  0.431876308
#> 94  094  0.332519589  0.955914442
#> 95  095 -1.237874527 -0.714051347
#> 96  096 -0.721189260 -1.261543523
#> 97  097  0.133639356 -0.230535124
#> 98  098  1.052192283 -0.964228731
#> 99  099  1.055351082  2.419544438
#> 100 100  0.704418793 -0.625589818

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.22303552 -0.885058221
#> 2   002  0.50392000 -1.551271013
#> 3   003  0.19141596  0.423388164
#> 4   004 -0.89096054 -0.475455935
#> 5   005 -0.05359765 -1.020701813
#> 6   006 -1.79551521 -1.659066372
#> 7   007  0.82519158 -0.408033897
#> 8   008 -1.47628434 -0.066283296
#> 9   009  0.20502379 -0.678907795
#> 10  010 -0.07843724  0.525060196
#> 11  011 -1.81071926 -0.478450745
#> 12  012  0.43946756 -0.536778181
#> 13  013 -0.05395872 -0.485401405
#> 14  014 -1.98410049  0.259733073
#> 15  015  1.03502980 -0.709045734
#> 16  016  0.58852282 -0.600497464
#> 17  017  0.05980696  1.139622880
#> 18  018  1.95653509  1.967916907
#> 19  019 -1.46687442  0.796331640
#> 20  020  0.90251769  0.865182517
#> 21  021 -0.07818234  0.027709878
#> 22  022 -0.14277345  0.410749939
#> 23  023  1.23357898  0.874492655
#> 24  024  1.27033624  1.766260058
#> 25  025  1.24202956  0.956084704
#> 26  026  0.14616473  1.349274760
#> 27  027 -0.46214607 -0.228125786
#> 28  028 -0.20631543 -0.398078184
#> 29  029  0.61441402  0.660484881
#> 30  030 -1.25701399 -1.747985327
#> 31  031 -1.40539038 -1.457886777
#> 32  032 -0.51769436 -0.775552291
#> 33  033  0.05824323 -0.021214787
#> 34  034  0.55789637  0.641343179
#> 35  035 -0.17270292 -1.870603460
#> 36  036 -1.41461098  0.011284785
#> 37  037  0.63489172  1.545266151
#> 38  038 -1.84867855 -1.177802924
#> 39  039  1.64079372 -0.530682928
#> 40  040  0.83145897  0.006602584
#> 41  041 -0.75787438 -0.774342870
#> 42  042 -0.21065708 -1.318071997
#> 43  043  1.77648709  0.028033905
#> 44  044  0.12163004 -0.463950099
#> 45  045  0.57203864 -0.827642259
#> 46  046 -1.17400783  0.580304082
#> 47  047  1.05875007  1.417011088
#> 48  048  1.48200433  0.560218009
#> 49  049  1.26656595  1.484709229
#> 50  050  1.75119627  0.655925274
#> 51  051  0.12597505  0.781029489
#> 52  052 -0.58035121  0.061441314
#> 53  053  1.06543830 -1.225829427
#> 54  054  0.26554867  0.189943319
#> 55  055 -0.12692619  0.344036779
#> 56  056 -0.19163467  0.036572410
#> 57  057  1.37502412  0.349680021
#> 58  058  1.44315776  2.200288125
#> 59  059  0.67931686 -2.022083718
#> 60  060  0.07929508  0.239979129
#> 61  061 -1.14246408 -0.925509649
#> 62  062  0.41924403  0.791036759
#> 63  063  0.01920156 -0.106329854
#> 64  064 -0.23350234  0.952692738
#> 65  065 -0.65920130 -0.342037003
#> 66  066 -0.23559642  0.396490654
#> 67  067  0.85315115  1.110602290
#> 68  068 -0.28281443 -0.988897726
#> 69  069 -0.97327700 -0.869005391
#> 70  070  0.09718919  0.053869858
#> 71  071 -0.58581769  0.309922991
#> 72  072 -0.49408151 -0.303531173
#> 73  073  0.19634118  0.426900148
#> 74  074 -0.13103988 -0.657542553
#> 75  075 -1.06905366  1.214794583
#> 76  076  0.03701070 -0.771102927
#> 77  077 -2.23349237 -0.001948052
#> 78  078 -1.16949829 -0.127328472
#> 79  079  1.89885246  0.288148466
#> 80  080  0.02820460  0.105036113
#> 81  081 -1.06303333  0.712730939
#> 82  082  0.55926553  0.529791118
#> 83  083 -1.67036227  0.370968572
#> 84  084  0.89367770  1.332682175
#> 85  085  1.82927151  1.591117848
#> 86  086  0.19808727 -0.550176566
#> 87  087 -1.38744327 -0.955412449
#> 88  088  0.19167636  0.104012891
#> 89  089  0.73334267  0.851551017
#> 90  090  0.06408908  0.898581007
#> 91  091 -0.13508268 -0.964291969
#> 92  092  2.02608981  0.095812824
#> 93  093 -0.89040979 -0.306336967
#> 94  094 -0.09744849 -1.198604989
#> 95  095  0.81391245 -0.485587652
#> 96  096 -0.47330178  0.245719542
#> 97  097 -1.13922740  0.175677392
#> 98  098  0.11167640 -0.024972867
#> 99  099  1.40540864 -0.942535943
#> 100 100  0.53575965 -1.038495757