Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.496499
#> Y_2 0.496499 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.8275474 -0.83667155
#> 2 -0.9716494 -0.59146808
#> 3  1.0452403  0.82507577
#> 4 -0.5860277  0.49853584
#> 5  0.3274854 -0.02148847

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.24203035 -0.916564990
#> 2   002 -0.60547550  0.490577853
#> 3   003 -1.12147886 -0.793967690
#> 4   004  0.44411287  0.449826329
#> 5   005 -1.04476932 -0.239801760
#> 6   006 -0.12899266  0.053390520
#> 7   007 -0.82984459 -0.305897235
#> 8   008  0.95608976  0.102553413
#> 9   009  0.10574086 -1.027650864
#> 10  010 -1.29687711 -1.331130482
#> 11  011  0.86077713 -0.193736945
#> 12  012 -1.72899145 -0.364984653
#> 13  013  0.15944006  0.725300383
#> 14  014  0.29630627 -0.668549499
#> 15  015 -0.87574158 -1.258043248
#> 16  016  1.56144131  1.200129720
#> 17  017  0.35865027 -2.142475170
#> 18  018 -0.52845040 -1.252431781
#> 19  019 -1.16471649 -1.219062686
#> 20  020 -1.57371108 -1.475308400
#> 21  021 -1.58379832 -1.828433669
#> 22  022  1.86954530  1.515825379
#> 23  023  0.40590318  1.206899687
#> 24  024  2.59361153  0.007152288
#> 25  025 -0.56002934 -0.404108088
#> 26  026 -0.04740891 -0.503501745
#> 27  027 -0.85695143 -0.797397718
#> 28  028 -1.30264245 -0.826231537
#> 29  029  0.94100381  0.810562533
#> 30  030  0.22520237 -0.050293491
#> 31  031  0.25874520  0.651474485
#> 32  032 -0.85509658  1.457728568
#> 33  033 -1.64842951  0.644706907
#> 34  034  0.32267874  1.648319287
#> 35  035  0.13350217 -1.313870822
#> 36  036 -0.93354367  0.534302851
#> 37  037 -1.51267806 -0.118704130
#> 38  038  1.86460698  1.278845075
#> 39  039  1.29493340  0.859220497
#> 40  040 -0.16522974 -1.598211591
#> 41  041  0.33198098  1.040133175
#> 42  042 -0.77124095 -0.168522715
#> 43  043  0.99255249 -1.623678406
#> 44  044 -0.03163218 -0.137097307
#> 45  045  1.06891357  0.744671694
#> 46  046 -0.81889329  0.090088307
#> 47  047  0.75952513  0.547310298
#> 48  048 -0.45922364 -0.661622361
#> 49  049 -1.14171952 -1.771047690
#> 50  050 -0.99265668  0.326765969
#> 51  051 -0.09138060 -2.112713333
#> 52  052 -0.95959993 -0.347038111
#> 53  053 -0.07821854 -1.337696567
#> 54  054  0.53827934 -0.848300609
#> 55  055 -0.57752709 -0.289537828
#> 56  056 -1.52235643 -0.760003721
#> 57  057  0.59797470  0.639904714
#> 58  058  0.11130432  0.269112544
#> 59  059 -0.60680737 -0.785681818
#> 60  060  0.75907294 -0.959495354
#> 61  061  0.01890621 -0.737325965
#> 62  062  0.32970318 -0.214675782
#> 63  063  0.02643427 -0.327175778
#> 64  064 -0.02016904 -0.232743593
#> 65  065 -0.12667995  0.722537945
#> 66  066  1.86951407  1.074374085
#> 67  067 -1.10938066 -0.829291739
#> 68  068 -0.15071771 -0.490028684
#> 69  069  2.17769342  0.565664748
#> 70  070 -0.28059046 -1.424211542
#> 71  071  0.40761012  0.538706359
#> 72  072  0.32273211  0.979403950
#> 73  073 -0.94053498 -1.771473848
#> 74  074 -2.47982057 -1.633358442
#> 75  075 -0.83990195 -0.760157517
#> 76  076  0.37470056  1.274214234
#> 77  077 -1.55073677 -1.376154483
#> 78  078 -0.15101721 -1.689837475
#> 79  079  0.57284840 -0.364257469
#> 80  080  0.59005644 -0.436799295
#> 81  081  0.65956265  1.094299019
#> 82  082 -1.05130265  0.790705149
#> 83  083 -2.33908825 -1.970721404
#> 84  084 -0.67773538 -0.205773362
#> 85  085  0.75067583  1.981722367
#> 86  086 -0.82828885 -1.061845356
#> 87  087  0.60982119  0.648572175
#> 88  088  0.15545612 -0.517316885
#> 89  089  0.99808989  1.244933207
#> 90  090 -0.74788724 -0.223053625
#> 91  091 -1.76031896 -1.024278587
#> 92  092  0.14318889 -0.501631432
#> 93  093 -1.40975854  0.209198360
#> 94  094 -1.24150913 -0.750911228
#> 95  095 -1.09182485 -0.663334565
#> 96  096  0.33426131  0.727028753
#> 97  097 -0.16949423  0.628268936
#> 98  098 -1.07502697  0.098603567
#> 99  099 -0.60526039 -0.676062843
#> 100 100  2.14584621  0.423692043

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.60835819 -0.114099019
#> 2   002 -0.90659246 -1.088362991
#> 3   003  0.29402454  0.193051565
#> 4   004 -0.34819653  0.845271577
#> 5   005 -0.96261686 -1.964434760
#> 6   006 -2.60515447 -1.829659079
#> 7   007 -0.17849621 -0.019802966
#> 8   008  0.88474600  0.875562690
#> 9   009  0.17432850  0.075315202
#> 10  010 -1.32793460  0.774837257
#> 11  011  1.02274064  0.367118342
#> 12  012  0.83218723 -0.593078639
#> 13  013  0.15628409  0.422234720
#> 14  014 -0.66472177 -1.587593622
#> 15  015 -0.84359255 -0.917371466
#> 16  016  0.58766517  0.474734884
#> 17  017  2.15620107  1.067745768
#> 18  018 -0.70210248  0.761406480
#> 19  019  1.03449134  0.040104146
#> 20  020  2.07470701 -0.026626527
#> 21  021 -1.43635388 -3.513983155
#> 22  022 -0.08205408 -1.571509226
#> 23  023 -1.31736495 -2.182158693
#> 24  024 -1.42141489 -0.404178853
#> 25  025  0.79340927  2.242116557
#> 26  026 -0.50540923 -1.144933256
#> 27  027  1.55174273  0.387266531
#> 28  028  1.37995703  0.484024404
#> 29  029 -0.59605743  1.026762435
#> 30  030  2.93350695  1.775824350
#> 31  031  1.23529155  1.745665785
#> 32  032  0.44222780  0.800107283
#> 33  033 -0.24123621  0.004139255
#> 34  034  0.17402355 -0.326527446
#> 35  035  0.73006494  1.355029666
#> 36  036 -1.83777105 -0.532929349
#> 37  037  0.16209210  0.865199229
#> 38  038 -1.75526144 -0.984672393
#> 39  039 -0.24523719  0.948370766
#> 40  040 -0.58964165 -1.012962015
#> 41  041  0.96155588  0.868089888
#> 42  042 -1.92089356 -0.858364566
#> 43  043  1.10024981  0.287712294
#> 44  044 -0.10638021  0.363963275
#> 45  045  0.65715484  0.143459488
#> 46  046 -0.90268935 -0.140259707
#> 47  047  0.09076395 -0.181188297
#> 48  048  0.65864619 -0.336487235
#> 49  049  0.06861009 -0.022501771
#> 50  050  1.49169989  0.514251828
#> 51  051 -0.07542474  0.009830981
#> 52  052 -0.06198382  0.770560273
#> 53  053  0.10996205  0.659986380
#> 54  054 -0.07984016  0.065705300
#> 55  055  0.60691410 -0.501894264
#> 56  056  0.64013843  0.081128626
#> 57  057 -1.21280742 -2.039451345
#> 58  058 -0.71239281 -0.541463236
#> 59  059  0.17969402  1.635097784
#> 60  060  1.22757788  1.704787668
#> 61  061 -0.69766617 -1.737745021
#> 62  062  0.60720679  0.297973314
#> 63  063 -0.22623657  0.093817474
#> 64  064 -1.06313149 -1.759281861
#> 65  065 -1.57526168 -0.267510625
#> 66  066  1.20488550  0.890271504
#> 67  067 -0.61895665 -0.109289858
#> 68  068 -0.48269502  0.995925774
#> 69  069  1.63600090  0.605363316
#> 70  070 -0.36045080  0.395376576
#> 71  071  1.42174649  1.243399881
#> 72  072 -0.21054559 -1.478250651
#> 73  073  0.13268047  0.674180507
#> 74  074  0.71455077  0.847210899
#> 75  075 -0.55034872 -1.113676228
#> 76  076 -0.15620173  0.491733399
#> 77  077  0.24127858 -0.659179947
#> 78  078  0.20250019 -0.403990228
#> 79  079 -0.01399923  1.691873588
#> 80  080 -1.24377438  1.744974668
#> 81  081  1.05904913  0.188586452
#> 82  082  0.36434067  1.094423006
#> 83  083 -1.04126760  1.150983690
#> 84  084 -0.61825889  0.930877185
#> 85  085  1.22277510  1.265801250
#> 86  086 -0.91869506 -0.534133410
#> 87  087  0.87121130  1.681750290
#> 88  088  0.37464784 -0.694431329
#> 89  089 -0.57439830 -0.880624399
#> 90  090  1.62766263  0.374911231
#> 91  091  0.93411783  0.623136307
#> 92  092 -0.63284005  1.346504655
#> 93  093 -0.89308046  0.757567415
#> 94  094 -2.25441062 -0.317894353
#> 95  095  0.27452752  1.289515735
#> 96  096 -0.85115094 -1.103195842
#> 97  097 -0.34713843 -0.882449728
#> 98  098 -0.37247042 -0.706267596
#> 99  099 -0.87775529 -1.319557327
#> 100 100 -0.59354396  0.112423409