Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4500941
#> Y_2 0.4500941 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.9718558  1.2783322
#> 2  0.4582700  0.1482993
#> 3 -1.4594664 -1.1623080
#> 4  0.6284573  0.2980785
#> 5  2.1026825  1.1550702

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  1.279593869  1.19861063
#> 2   002  1.304314319  0.32046855
#> 3   003  1.093883016  1.38406422
#> 4   004 -0.351495447  0.07420189
#> 5   005  1.770788470  0.87913755
#> 6   006 -1.402421031 -0.16970387
#> 7   007  1.666100350 -0.52757628
#> 8   008 -0.256956778  0.30328892
#> 9   009 -0.265268423  0.23521808
#> 10  010 -1.472892023 -0.84465246
#> 11  011  0.132030788  1.12461333
#> 12  012  0.465150986  1.30601903
#> 13  013 -1.826958444 -2.25296398
#> 14  014 -2.490875741 -1.73351752
#> 15  015  2.032627708  0.65527230
#> 16  016  0.600834583  0.95901038
#> 17  017 -0.149064944  0.30869440
#> 18  018 -0.584266318  0.10614914
#> 19  019 -0.971460331 -0.96556500
#> 20  020 -1.192672620 -1.46266422
#> 21  021 -0.437528685 -0.50839413
#> 22  022 -1.014921341  0.22849092
#> 23  023 -1.559408001  0.39965133
#> 24  024 -0.448531666  0.05055767
#> 25  025 -0.300952738 -1.49708647
#> 26  026 -0.390546777 -0.46829135
#> 27  027  1.392492051  0.45016737
#> 28  028 -0.603134899 -0.73484705
#> 29  029  2.208691740  1.09794399
#> 30  030  0.075545526  1.55312523
#> 31  031  0.754391409  0.14577373
#> 32  032 -0.817147344  0.20371162
#> 33  033 -0.154948322 -0.46346412
#> 34  034 -0.459697802 -0.29279518
#> 35  035 -0.298852076 -1.44478026
#> 36  036  1.044504285  0.40989869
#> 37  037 -0.377273857 -0.16599215
#> 38  038 -1.044677596 -1.62105353
#> 39  039  1.189151486 -1.00409443
#> 40  040  0.577849349 -2.00597350
#> 41  041 -0.638898694  0.61626175
#> 42  042  0.396801221  0.83701032
#> 43  043  0.452966747 -0.41592725
#> 44  044  0.098747243  0.05637417
#> 45  045  0.704885265  0.84958463
#> 46  046  0.289578893  0.09710725
#> 47  047 -0.409808179 -1.56982863
#> 48  048  0.603166511  0.01002267
#> 49  049  0.453642765 -0.39511393
#> 50  050 -0.016610257  0.70110175
#> 51  051 -0.042206349  0.06860973
#> 52  052 -0.774147193 -0.74813583
#> 53  053  0.909524231  1.60792009
#> 54  054 -0.427489180 -0.86046547
#> 55  055  1.066353290  1.67337350
#> 56  056  0.471827765  0.75392548
#> 57  057 -0.644348653 -1.36483549
#> 58  058  0.646757493  0.41384736
#> 59  059 -1.670923959  0.30123356
#> 60  060 -0.025444794 -0.97702650
#> 61  061 -0.509261218  1.26933300
#> 62  062 -0.498036764 -1.12801948
#> 63  063  0.432306305  0.56271004
#> 64  064  0.892250497  1.60781779
#> 65  065  0.353995741 -0.58800452
#> 66  066 -0.231374198 -0.08458433
#> 67  067 -1.107308088 -0.59186707
#> 68  068 -0.435072444 -1.16529579
#> 69  069 -0.373274313 -1.04244382
#> 70  070 -1.738469431 -2.26848503
#> 71  071 -0.842287000  0.38166454
#> 72  072  2.061562002  0.93871264
#> 73  073 -0.048859151  0.27011037
#> 74  074  0.716086132  0.28083172
#> 75  075  0.273142385  1.83632831
#> 76  076  0.052595926  0.69892184
#> 77  077 -0.266663920  0.82572772
#> 78  078  1.307516765 -2.14068336
#> 79  079  0.212427559  0.56954654
#> 80  080  2.456227362  1.82166897
#> 81  081  1.351633190 -0.55247594
#> 82  082  0.565200465 -0.09461621
#> 83  083  0.409472716  0.27232984
#> 84  084  1.394940189 -0.55108329
#> 85  085  0.005161736  0.29241721
#> 86  086 -0.933893466 -0.16201764
#> 87  087 -0.333289394  0.44018291
#> 88  088  0.186773502  0.32274712
#> 89  089 -1.055910219  0.19686873
#> 90  090  0.042692056  0.18002251
#> 91  091  1.478989696  0.66071475
#> 92  092  0.939572433 -0.86887933
#> 93  093 -2.223087727 -2.02879467
#> 94  094 -1.602212344  1.42806626
#> 95  095  0.229151825  1.42477799
#> 96  096  1.016436936  0.06105886
#> 97  097 -0.191891659 -1.22217554
#> 98  098 -2.167455484 -0.02059626
#> 99  099 -2.286640623 -1.86011600
#> 100 100  0.278366197 -0.44939319

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -1.17540261 -8.212427e-01
#> 2   002  1.10998035 -1.213384e+00
#> 3   003  1.74993896  9.329907e-01
#> 4   004 -0.55822300  1.102139e+00
#> 5   005  0.63286774  9.218165e-01
#> 6   006  1.76248769  9.767572e-01
#> 7   007  1.24616496 -8.619076e-06
#> 8   008  0.47406707  5.284315e-01
#> 9   009  0.37820133 -6.533302e-01
#> 10  010 -0.85405225 -2.135113e+00
#> 11  011 -0.01391353 -1.254969e-01
#> 12  012 -0.94224775 -1.946816e+00
#> 13  013 -0.96916876 -1.524689e-02
#> 14  014  0.15275320 -4.835297e-01
#> 15  015  0.24572097  1.282604e+00
#> 16  016 -1.94852891 -1.688983e+00
#> 17  017  1.10061562  8.443620e-01
#> 18  018  2.21197899  1.069418e+00
#> 19  019 -2.08551445  3.711067e-02
#> 20  020 -0.93679229 -1.786484e-01
#> 21  021  0.65475135  3.128734e-01
#> 22  022 -0.63769014 -3.153231e-01
#> 23  023  0.38118647 -1.502492e-01
#> 24  024 -0.24491141  5.247869e-01
#> 25  025 -1.26387648 -9.171708e-01
#> 26  026  1.41494950  8.629275e-01
#> 27  027  0.27392730 -3.526834e-01
#> 28  028 -0.14928207 -6.925635e-01
#> 29  029 -0.25385949 -3.663798e-01
#> 30  030 -1.43743686 -1.449843e+00
#> 31  031  1.13945347  1.028436e+00
#> 32  032  1.31993573  5.963135e-01
#> 33  033 -0.13336173  6.554666e-01
#> 34  034  0.13725200 -1.718491e+00
#> 35  035 -0.51416413 -2.560958e-01
#> 36  036 -1.38578013 -5.903967e-01
#> 37  037 -1.87309459  1.304444e-03
#> 38  038  0.97782233 -9.532745e-02
#> 39  039  2.19781990  1.610425e+00
#> 40  040  1.15990130  1.890325e+00
#> 41  041 -0.45792323 -2.957525e-01
#> 42  042 -0.03077135  1.870013e+00
#> 43  043  1.99121125 -3.602679e-01
#> 44  044  2.28420607  3.054961e-01
#> 45  045 -1.17024615 -9.035650e-01
#> 46  046  0.41755327  1.497240e+00
#> 47  047 -0.56623204 -3.563768e-01
#> 48  048  0.62375951  1.198607e+00
#> 49  049  1.03104126  2.053617e-01
#> 50  050  1.17433216  6.308501e-01
#> 51  051  0.40552344 -9.865007e-01
#> 52  052  0.43529461 -4.601216e-01
#> 53  053  0.52373310  5.450389e-01
#> 54  054 -1.35738121 -1.575618e+00
#> 55  055  0.52132610 -7.591199e-01
#> 56  056 -0.63577494 -1.220509e+00
#> 57  057 -1.29972168 -2.021189e+00
#> 58  058  1.61971858  9.906872e-01
#> 59  059 -1.16419789  2.494284e-01
#> 60  060  0.19451665 -7.590474e-01
#> 61  061  0.47617748  1.831054e-01
#> 62  062  1.77241587  7.896070e-01
#> 63  063 -1.70729878 -1.143364e+00
#> 64  064  0.37401375  8.585957e-01
#> 65  065  0.79983313  5.048515e-01
#> 66  066 -1.58965696 -1.011727e+00
#> 67  067 -0.47306306 -1.004246e-01
#> 68  068  1.01081025  4.871323e-01
#> 69  069  0.88765197 -2.425727e-01
#> 70  070  0.14312113  5.184491e-01
#> 71  071  0.04849063  1.052930e-01
#> 72  072 -1.17757642 -5.038297e-01
#> 73  073 -0.05349788  9.189644e-01
#> 74  074  0.49666801  8.293285e-01
#> 75  075 -0.35393036 -1.747848e+00
#> 76  076  0.64879994  1.114207e+00
#> 77  077 -1.03275553 -1.117538e+00
#> 78  078  0.02349342 -4.578238e-01
#> 79  079  0.99104195  1.390999e+00
#> 80  080 -0.25378462 -4.442814e-01
#> 81  081  0.42079218 -1.264557e-01
#> 82  082  0.46526559  1.223753e+00
#> 83  083 -0.41565733 -1.319832e+00
#> 84  084  1.62218120 -3.096804e-01
#> 85  085  1.31474696 -4.565849e-02
#> 86  086  1.68672171  2.294748e+00
#> 87  087  1.48362958  6.254475e-01
#> 88  088 -1.02218562 -8.882439e-01
#> 89  089 -0.19195965 -7.465433e-01
#> 90  090  0.27815756  9.162390e-01
#> 91  091 -0.40593842 -9.675482e-01
#> 92  092  0.09612265 -1.411423e+00
#> 93  093 -1.02047983 -1.469713e+00
#> 94  094  0.45289727 -4.347518e-01
#> 95  095  0.50114794  5.320766e-01
#> 96  096  0.23978308  1.093097e+00
#> 97  097  1.73098963 -1.404725e-01
#> 98  098 -0.70633080  7.854913e-01
#> 99  099 -0.82151761  7.905780e-01
#> 100 100  0.38848725 -1.702831e-01