Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4240759
#> Y_2 0.4240759 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  0.30690371  1.4668971
#> 2  1.40152945  1.8314931
#> 3 -0.97884654 -0.7787543
#> 4  0.65954236 -0.6457188
#> 5  0.02535303  0.0632930

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.485864625  1.309960402
#> 2   002  0.343858024  0.054088956
#> 3   003  0.013216820  0.682494958
#> 4   004 -0.702358774  1.431899040
#> 5   005 -0.472086642 -0.924233791
#> 6   006 -0.911668818 -0.065721323
#> 7   007 -1.809188487  0.011518844
#> 8   008 -0.077721553 -0.886264894
#> 9   009  0.282879452  0.962087958
#> 10  010  0.197691783 -1.344504308
#> 11  011  0.744312983 -0.066222180
#> 12  012  0.544255053 -0.543359400
#> 13  013  0.189432809 -0.673145399
#> 14  014 -0.343120816 -0.660164061
#> 15  015 -0.062020226  0.974485994
#> 16  016  0.002213026 -0.316162834
#> 17  017 -0.334382623 -1.145715271
#> 18  018  0.061739090  0.017275760
#> 19  019  1.337258621  0.267474035
#> 20  020 -0.296691772  0.671391788
#> 21  021  1.069919755  0.712273355
#> 22  022  0.906939102  0.521400687
#> 23  023  0.917907525 -0.453208141
#> 24  024  0.863722001  0.788246434
#> 25  025 -0.970670255  0.586341062
#> 26  026 -0.613089051 -1.062209496
#> 27  027  3.551264813  1.680873045
#> 28  028  1.440161828 -0.302123554
#> 29  029  2.552318281  0.218088870
#> 30  030 -1.378605302 -1.826254241
#> 31  031  1.050477796 -0.245547062
#> 32  032  0.710135152  0.800885870
#> 33  033 -0.316445962 -1.369265259
#> 34  034  0.826448232  0.989684138
#> 35  035 -1.259123328 -0.603400174
#> 36  036 -1.431874692  0.934958052
#> 37  037 -0.298298282 -1.218994616
#> 38  038  1.730871680 -0.993435565
#> 39  039  1.895647585 -0.111450634
#> 40  040  2.649478114  0.567796149
#> 41  041 -0.602426267 -0.111450682
#> 42  042 -0.630797750 -0.312833685
#> 43  043 -1.974909101 -1.292382247
#> 44  044 -0.205183794 -0.223278341
#> 45  045  0.829401315  1.118818023
#> 46  046  0.403631880 -1.938133648
#> 47  047 -1.534738703 -1.321447112
#> 48  048  2.864429551  1.384716412
#> 49  049 -1.744739233  0.362311834
#> 50  050 -1.498796238 -1.121472676
#> 51  051 -0.777738261 -0.512545671
#> 52  052 -1.777011825 -0.905356251
#> 53  053  0.090498420 -0.345414260
#> 54  054  0.355400306 -1.166018953
#> 55  055 -0.413795033 -0.995982694
#> 56  056  0.269133608 -0.016236039
#> 57  057  0.222981960  0.383458907
#> 58  058  0.900281619  0.531354400
#> 59  059  1.711300939  0.618530021
#> 60  060  0.387367691 -0.631468001
#> 61  061 -0.364157373  0.014394870
#> 62  062  0.044590966 -0.274291306
#> 63  063  0.259916912 -0.165918806
#> 64  064 -0.399233942 -0.328118319
#> 65  065 -0.873591370  0.163696887
#> 66  066 -1.396619493  0.052072646
#> 67  067  0.788824029  1.687583731
#> 68  068  0.581115852  0.474825099
#> 69  069  0.007915756  1.294659343
#> 70  070  0.058914741  0.578186106
#> 71  071 -0.828099001 -0.496500244
#> 72  072 -0.823963297  0.306528319
#> 73  073  0.246502690 -0.818385936
#> 74  074 -0.534271809  0.532870184
#> 75  075  0.035654325  0.990281026
#> 76  076 -0.163119135 -2.166228736
#> 77  077  0.491766192  1.242428962
#> 78  078 -0.559325860  0.201265680
#> 79  079  0.767115720  1.624759600
#> 80  080  0.446071709  0.184257711
#> 81  081  0.023672547  0.849929739
#> 82  082 -0.575567490 -0.289199646
#> 83  083  0.391991904 -0.934136664
#> 84  084  1.089687200 -1.226052504
#> 85  085  0.881843742 -0.686292486
#> 86  086  2.163489679  1.719474069
#> 87  087  0.998360354  0.003865948
#> 88  088 -0.608970893 -0.467288703
#> 89  089 -0.049817602 -1.825654782
#> 90  090  0.294697679 -0.617820002
#> 91  091 -1.326784551  1.499037030
#> 92  092  0.280712278 -1.014751457
#> 93  093 -0.155376677 -0.562063419
#> 94  094  2.068786473  0.680114652
#> 95  095  0.609621176  0.356622252
#> 96  096 -0.602414893 -1.085460663
#> 97  097  0.382307830 -0.674695610
#> 98  098 -0.581132899  0.513301932
#> 99  099  0.339684407 -0.053036730
#> 100 100  0.255219517  0.335577939

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1           Y.2
#> 1   001  0.118829054 -0.9956025183
#> 2   002  2.276543307  0.4370522276
#> 3   003 -1.474494663 -0.0119411054
#> 4   004 -1.098426733 -0.8686323136
#> 5   005 -0.011271698  0.3555712217
#> 6   006 -0.018454449 -0.4561336794
#> 7   007  0.452973156  0.3519016765
#> 8   008  0.137176889  0.6176151617
#> 9   009 -0.099875989 -0.8075507040
#> 10  010  0.009571143  1.4397222221
#> 11  011 -0.372438843 -0.4896494560
#> 12  012  0.616031242  0.5345161954
#> 13  013  0.013830237 -1.1533059491
#> 14  014 -0.167109479  1.3519369682
#> 15  015 -2.217296382 -0.0007877873
#> 16  016 -1.637850738 -1.1910916837
#> 17  017 -1.343873919 -0.0998347534
#> 18  018  0.183704557  0.4448790183
#> 19  019  0.888872275 -0.2035487420
#> 20  020  0.973557132 -0.9809088951
#> 21  021 -0.977208308 -0.8573610062
#> 22  022  0.337337500 -0.1937536947
#> 23  023  1.248399449 -0.9664335016
#> 24  024  0.735172228  1.6135092857
#> 25  025 -0.592288869 -0.1992814673
#> 26  026  0.308282684 -0.4863752595
#> 27  027 -0.469905930  0.1003824000
#> 28  028 -2.437616687 -2.3124780838
#> 29  029  0.502404931  0.5291823679
#> 30  030 -0.810766821 -0.9775500307
#> 31  031 -0.429395028 -1.0428371729
#> 32  032  0.214050790 -0.2905282632
#> 33  033 -0.316206778 -0.3125171532
#> 34  034  1.026541147 -0.0545268785
#> 35  035  0.228240550  0.2282584980
#> 36  036 -2.344445539 -0.6100548522
#> 37  037 -1.311553640 -0.6678384854
#> 38  038  0.760479757  0.6984040895
#> 39  039 -1.003197046  0.6100582418
#> 40  040 -0.813056014 -0.0409531995
#> 41  041  0.143289677 -0.0188545230
#> 42  042 -0.411764468  0.6602691248
#> 43  043  1.053759193  1.9491129264
#> 44  044  0.533195079  0.4674611820
#> 45  045 -0.016587387  0.5185406265
#> 46  046 -0.589960893 -1.1265709162
#> 47  047 -0.981585070 -0.1759470410
#> 48  048 -0.806053526  0.4029894430
#> 49  049  2.073671690  2.0224885601
#> 50  050  0.102239446  0.0002442047
#> 51  051 -1.863426000 -1.8622030305
#> 52  052  0.802819631  0.2286160613
#> 53  053  0.168583962  0.9691014559
#> 54  054 -0.541087246 -0.9352338499
#> 55  055 -1.801360346  0.0248939954
#> 56  056 -0.575471007 -1.0972099554
#> 57  057 -1.623384615  0.4554988972
#> 58  058 -1.903011564  0.8858874941
#> 59  059 -1.327285356  1.7456881229
#> 60  060  1.217030599  1.4454507074
#> 61  061  0.369109965 -0.2217709204
#> 62  062 -1.208413421  0.8617745846
#> 63  063  1.634153345 -0.8357775732
#> 64  064  0.814559503  1.1186202898
#> 65  065  0.822329760 -1.5005744426
#> 66  066  0.289287953  0.0699276396
#> 67  067  1.287224948  0.5574874732
#> 68  068 -0.094055526 -0.1058624052
#> 69  069 -1.278468329 -1.3649332879
#> 70  070  0.310146049 -0.3856438650
#> 71  071 -0.227378869 -0.6248810355
#> 72  072  0.106077599  1.0279625873
#> 73  073 -1.142998333 -0.8718598884
#> 74  074 -0.294434768  0.6284016099
#> 75  075 -1.330437112 -0.4402358004
#> 76  076  0.464763671 -0.1854025132
#> 77  077 -0.181221817 -1.0025622312
#> 78  078  1.026593168 -0.7451441422
#> 79  079  0.344248728 -1.1119897263
#> 80  080  0.922002079 -0.1844607610
#> 81  081  0.020979096  0.5172923360
#> 82  082  0.144759146 -0.2527608773
#> 83  083 -0.387791986 -2.2581541477
#> 84  084 -0.122050078 -0.3552683154
#> 85  085 -1.187787951 -1.3930460849
#> 86  086  0.286245499  0.3782800533
#> 87  087 -0.404220023 -0.6420434079
#> 88  088 -0.166228066  0.7239442743
#> 89  089 -0.703866641 -1.1681839323
#> 90  090 -0.197759744 -0.6309082301
#> 91  091  1.949051501 -0.3581567384
#> 92  092  1.022995415 -0.0901400836
#> 93  093 -1.520248544 -0.4860391546
#> 94  094 -0.223557197  0.5520479616
#> 95  095  1.530130626  1.1169727837
#> 96  096 -0.756517413  0.2326237662
#> 97  097 -0.412680182 -1.5423730962
#> 98  098 -0.258081238  0.9504942673
#> 99  099 -1.030530769 -1.1336489536
#> 100 100 -0.153574559  0.9015330647