Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5455032
#> Y_2 0.5455032 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.5532629  1.1014768
#> 2  1.5852818  0.3416641
#> 3 -0.6109166  0.4885296
#> 4 -0.8931749 -0.7286858
#> 5  1.0629278 -0.2624164

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -1.611928448 -0.76832337
#> 2   002  0.772780350 -0.42474201
#> 3   003 -0.089127681 -0.42808455
#> 4   004 -0.573999094  0.59002420
#> 5   005 -0.982617471  0.53106975
#> 6   006  1.221011222  1.04908203
#> 7   007  0.955760648 -0.55434026
#> 8   008  0.011281971 -0.14224794
#> 9   009  0.706514030  0.23250775
#> 10  010 -0.660168755 -0.48326144
#> 11  011  0.652609670  2.10171657
#> 12  012 -0.293918098  0.38303624
#> 13  013 -0.106663702 -0.67525921
#> 14  014 -0.507450390  0.03907045
#> 15  015 -1.373662254 -0.16892561
#> 16  016 -0.298290010  0.54663056
#> 17  017  0.360439553  0.39594890
#> 18  018  0.566380073  0.36892325
#> 19  019 -0.184230880 -0.04240436
#> 20  020 -1.247306260 -1.65348022
#> 21  021 -0.915862534 -0.65020574
#> 22  022  1.459912199 -0.48833617
#> 23  023 -0.009209168  0.67350267
#> 24  024 -1.299556231 -1.92657804
#> 25  025  1.206474851  1.61171542
#> 26  026  0.615652130  0.23050222
#> 27  027 -2.034704520  0.17778889
#> 28  028  0.621059686 -0.42016762
#> 29  029  0.727238480  0.26568415
#> 30  030 -0.719844834  0.03044011
#> 31  031 -1.505690665 -1.29079118
#> 32  032 -1.074994546  0.01633940
#> 33  033 -0.021975802  2.00789116
#> 34  034  1.491887882 -0.22527739
#> 35  035 -0.787783465 -0.13358584
#> 36  036  2.098334323  0.15183885
#> 37  037  0.079420130 -0.61597977
#> 38  038 -0.967133316 -0.72078437
#> 39  039 -0.801568014 -0.01532617
#> 40  040  0.575322612  1.00904278
#> 41  041  0.362599094 -1.06369849
#> 42  042  0.133459527 -0.15091875
#> 43  043  0.624197903  1.01618910
#> 44  044  0.202394069 -0.87375319
#> 45  045 -0.327436743 -0.54366879
#> 46  046  0.590740706 -0.44101158
#> 47  047 -0.464952652 -1.14984155
#> 48  048 -1.730327917 -1.65138482
#> 49  049  0.677682791  0.42270514
#> 50  050 -1.857270887 -0.07719619
#> 51  051  0.896480405 -0.54529134
#> 52  052 -0.412693139  1.49626731
#> 53  053  2.017538459  0.69360907
#> 54  054  1.684739425 -1.53633029
#> 55  055 -0.647069359 -0.35384727
#> 56  056 -0.054641017  0.16411347
#> 57  057  0.523885867 -0.50572082
#> 58  058 -0.002860622 -0.47753642
#> 59  059  0.226760870  0.44113134
#> 60  060 -1.097743742 -1.21490241
#> 61  061 -0.031510037 -0.49450163
#> 62  062  0.524812950 -0.80305230
#> 63  063 -0.802957329 -0.55594249
#> 64  064  0.962820739  0.43334390
#> 65  065  0.526881302  0.50928577
#> 66  066 -0.965104061  0.29775675
#> 67  067  0.227377572  1.72357044
#> 68  068  0.630998764  0.05622114
#> 69  069 -0.967304409 -0.37415918
#> 70  070 -0.235050104  0.68609103
#> 71  071  1.337727074 -0.21330475
#> 72  072 -0.678140031 -1.51568384
#> 73  073  0.257629930 -0.22651423
#> 74  074  0.770114137  0.82213036
#> 75  075  0.501796082 -0.42983697
#> 76  076  0.530528063  0.52785700
#> 77  077 -0.729896159  1.57854930
#> 78  078  0.081830699 -0.24950984
#> 79  079  1.592194162  1.34456519
#> 80  080  0.276916563  0.34565427
#> 81  081  1.395725992  0.55740068
#> 82  082  1.217187275  2.02396025
#> 83  083 -1.368244531 -0.91012217
#> 84  084 -0.538147224  0.99616837
#> 85  085  0.126287037  0.03965746
#> 86  086 -0.578340232  0.13078816
#> 87  087 -1.010113581 -0.12539466
#> 88  088 -1.087559824 -0.10851810
#> 89  089  0.955411221  0.52167636
#> 90  090  0.097150138 -0.67209823
#> 91  091  0.229007977  0.13908730
#> 92  092 -2.385218137 -1.56673034
#> 93  093 -0.200267967 -1.00460145
#> 94  094 -1.449860101 -0.10915729
#> 95  095 -0.355270928  0.60188653
#> 96  096 -1.248552858  0.05612192
#> 97  097  1.351879761  1.06193653
#> 98  098  0.841429643  0.90902712
#> 99  099 -1.519424204 -0.60112883
#> 100 100 -1.143034244 -0.33181838

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001  1.36725419  0.8711549468
#> 2   002 -1.31743079 -1.1863717921
#> 3   003  0.52592124  0.8641212626
#> 4   004  0.07239961  0.0380859726
#> 5   005 -0.53488212  1.1903329177
#> 6   006 -0.16087519 -0.3741363389
#> 7   007  0.70787430  0.5181607649
#> 8   008  0.13773243  0.0492948650
#> 9   009  0.22033882  1.6669358397
#> 10  010 -0.71716839 -0.4219860956
#> 11  011 -0.59208922 -0.7846553614
#> 12  012  1.42542341  0.7270575068
#> 13  013 -1.42288901 -1.4272454736
#> 14  014  1.99888617  0.2944291689
#> 15  015 -0.73267920 -1.6412796838
#> 16  016 -0.36956460 -0.1889003950
#> 17  017 -0.53345368 -1.5488108443
#> 18  018 -1.00040811 -1.2382098937
#> 19  019 -0.28778594  0.1043342749
#> 20  020  0.24163057 -0.3906237632
#> 21  021 -0.12995683  0.8877754220
#> 22  022 -1.15928323 -0.0558090411
#> 23  023 -1.06128235 -0.7541248451
#> 24  024 -0.82957592 -0.0996743575
#> 25  025  0.79973093  1.2613793652
#> 26  026 -1.10484792 -0.0976388532
#> 27  027 -1.53322069 -1.3111370471
#> 28  028 -0.45474798 -1.7526116689
#> 29  029 -0.83052434 -1.0244816242
#> 30  030 -0.92235060 -0.5507708347
#> 31  031 -2.02378656 -1.5961268770
#> 32  032 -0.64933957 -0.7764078263
#> 33  033 -0.58464891 -0.8563622560
#> 34  034  0.82321723 -0.1431378014
#> 35  035  0.96763787  0.5587695249
#> 36  036  0.45092701 -0.2705321943
#> 37  037 -1.73369759 -0.9058812679
#> 38  038 -1.26710057 -1.1595471933
#> 39  039 -0.87269348 -0.2836628678
#> 40  040  1.47467305  0.5706297436
#> 41  041  0.25242136 -0.0730477771
#> 42  042  0.78281894 -0.0009178267
#> 43  043 -0.77704820 -0.6065130049
#> 44  044  1.85867868  0.4277915679
#> 45  045 -0.61196659  1.6781935210
#> 46  046  0.45084452 -0.4275859566
#> 47  047  1.26952799  0.2288960785
#> 48  048 -1.41867255 -0.6688721037
#> 49  049  0.84988819  0.6214936905
#> 50  050  1.10848689  0.2185730597
#> 51  051  0.36326945 -0.3204578738
#> 52  052  0.02094898  0.5479793588
#> 53  053  0.31182602  0.2805398881
#> 54  054  0.47041812  2.7241594271
#> 55  055 -0.40081831  1.6306500413
#> 56  056 -0.79122202 -0.8957813178
#> 57  057  0.28567533  0.5486501204
#> 58  058 -0.31151312 -0.1945328633
#> 59  059  1.74904216 -0.5141630776
#> 60  060 -0.68542982 -1.4746354520
#> 61  061 -0.65500368 -0.1412486204
#> 62  062  0.67131999  0.3539952089
#> 63  063  1.18563433  0.9968026036
#> 64  064  0.44783831 -1.1024353379
#> 65  065 -0.06940195 -0.0261373325
#> 66  066  0.72601334  0.0665655700
#> 67  067  0.52502250  0.6204289282
#> 68  068 -0.13998761 -0.4888910313
#> 69  069 -1.80718928 -0.5699794463
#> 70  070  1.38882970  0.8848709288
#> 71  071 -0.93600958 -1.5100036425
#> 72  072 -1.02290301  0.5986572204
#> 73  073  0.05778350  0.8801201770
#> 74  074 -0.96562166 -0.0006805301
#> 75  075  1.07220811 -0.0404040143
#> 76  076 -0.59520133 -1.1696671944
#> 77  077 -0.59809746  0.4961719546
#> 78  078 -0.24121261 -0.6834514163
#> 79  079 -0.62158555  0.8907720503
#> 80  080 -0.32008656 -0.7503192347
#> 81  081 -0.08617582  1.0651799606
#> 82  082 -0.88765298  0.4466585348
#> 83  083  0.14552820 -0.8201727320
#> 84  084  0.28711528 -1.8259799596
#> 85  085  0.62524098  0.0942371274
#> 86  086  0.97909197  1.5104637894
#> 87  087  0.26883356 -0.2719698900
#> 88  088 -1.84683675 -0.5774684511
#> 89  089  0.74929484  0.9028278555
#> 90  090 -0.04501012 -0.1774185507
#> 91  091 -0.02650313  0.1740715356
#> 92  092  1.11999841  1.7002591911
#> 93  093 -0.48153734  0.5990229365
#> 94  094 -1.55333871  0.0142938122
#> 95  095 -2.26014665 -0.7798810985
#> 96  096  0.07292878  0.0171086716
#> 97  097 -1.38769165  0.0494485358
#> 98  098  1.46798831  0.9923654099
#> 99  099  1.94625723  1.2277291530
#> 100 100  0.10464832 -0.6290243453