Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.513425
#> Y_2 0.513425 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1         Y_2
#> 1  0.17387328  0.43897269
#> 2  0.18529065 -1.07208289
#> 3 -0.87623782 -0.02600521
#> 4 -0.04551395 -1.43720370
#> 5  0.44101859  0.37412403

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.66966380  1.19652269
#> 2   002 -2.46359802 -0.75906118
#> 3   003 -0.07380473 -0.21119165
#> 4   004  0.94506136  0.33064636
#> 5   005  0.89810677 -0.59171251
#> 6   006  1.43206612 -0.37975567
#> 7   007 -2.33430094 -0.78631307
#> 8   008 -0.86063843 -0.62232445
#> 9   009 -0.07737719  0.82346809
#> 10  010 -0.03720283 -0.37335445
#> 11  011 -0.74442328 -0.32762739
#> 12  012  0.23802934 -0.23543021
#> 13  013 -0.33478171  0.06339394
#> 14  014 -0.36063001 -0.33880858
#> 15  015  0.19546115 -1.24583004
#> 16  016  0.51703323 -0.25696611
#> 17  017 -0.22241445 -0.06630168
#> 18  018  0.50889532  1.03888304
#> 19  019 -2.96810823 -1.39871762
#> 20  020  0.28658124 -0.85936528
#> 21  021 -0.16385620 -0.04261278
#> 22  022  0.72927759 -1.21299531
#> 23  023 -1.33474507 -1.76188800
#> 24  024 -0.27530415  0.96909357
#> 25  025  0.51572737  0.16204557
#> 26  026  1.72685975  3.21094630
#> 27  027  1.82974268  0.85238400
#> 28  028  0.94778486  0.24221577
#> 29  029  0.14277180  0.40776912
#> 30  030  0.99997817 -0.90435567
#> 31  031  0.70519996 -0.08332531
#> 32  032  0.95464566 -0.42016071
#> 33  033  0.71046776 -0.68182427
#> 34  034  1.33539433  0.41199143
#> 35  035 -0.75653062 -0.18569014
#> 36  036 -0.41534486 -0.41591460
#> 37  037 -0.02275048  0.38235949
#> 38  038 -0.14045526 -0.63656575
#> 39  039 -0.03589785 -0.57220808
#> 40  040  0.94946864 -0.38653422
#> 41  041 -1.33943100 -0.89102639
#> 42  042  0.79632040  0.91897145
#> 43  043  0.59059256  1.16332373
#> 44  044 -0.69623071 -1.79007888
#> 45  045 -1.60675587 -1.31687375
#> 46  046  1.48051638  1.04679907
#> 47  047  0.73862769  0.86776215
#> 48  048  0.25113729 -0.58738420
#> 49  049 -0.61970837 -0.94873325
#> 50  050 -1.46490236  0.33055820
#> 51  051 -0.61677222 -0.87005013
#> 52  052  0.19586330 -0.20958414
#> 53  053 -0.99801937 -0.17467931
#> 54  054 -0.52131067  0.19871668
#> 55  055  2.26135593  0.57862105
#> 56  056  0.41268157  0.49035977
#> 57  057 -2.56076321  0.02133361
#> 58  058  0.59920599 -0.64259535
#> 59  059 -0.63874623 -0.50935175
#> 60  060 -0.62361404 -0.30660393
#> 61  061  0.22531150 -0.79519530
#> 62  062 -0.12897706 -0.81366156
#> 63  063  1.00321490  0.20549379
#> 64  064  0.54942788 -0.81635376
#> 65  065  0.41954588  1.08124830
#> 66  066  0.40321506  0.53417950
#> 67  067 -0.28495455  0.93225249
#> 68  068 -0.61092755 -1.48072807
#> 69  069 -1.30862124 -0.20570912
#> 70  070 -1.42516701 -1.48496288
#> 71  071 -0.87839013 -1.52084806
#> 72  072 -1.82835475 -0.82736153
#> 73  073 -0.88036316 -1.00814915
#> 74  074 -0.70580897  0.09617806
#> 75  075 -0.42088569  0.33073603
#> 76  076  0.44127960  0.22155858
#> 77  077 -1.10431894 -0.86000442
#> 78  078  0.46175293  0.58072836
#> 79  079 -1.49092484 -0.25011194
#> 80  080  0.26242448  0.36149091
#> 81  081  1.32389328 -0.23042665
#> 82  082 -0.90263939 -0.44846321
#> 83  083  1.01824260  0.50804788
#> 84  084  0.61538639  1.23647810
#> 85  085  1.18899643 -0.65300410
#> 86  086  0.18000168 -0.77596220
#> 87  087 -0.47543827  1.61371381
#> 88  088  0.13396639 -1.40521068
#> 89  089  0.50003124 -0.50096369
#> 90  090  0.90629133  1.31201449
#> 91  091  0.09574607  0.52366137
#> 92  092  0.63549790  1.05008994
#> 93  093  1.03598491  0.20624333
#> 94  094 -1.32119861  0.10109991
#> 95  095 -0.75052245 -0.66184319
#> 96  096  0.07036282  0.11735982
#> 97  097 -0.19124879  1.60544420
#> 98  098  2.06499549  0.30335923
#> 99  099 -0.41633290 -1.27033909
#> 100 100  0.33991290 -0.03978608

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.457868851  1.21147168
#> 2   002 -1.429198499 -1.24696232
#> 3   003 -0.843411984  1.06503918
#> 4   004 -0.639091903 -0.22444024
#> 5   005  0.511883492  1.17988116
#> 6   006 -0.192591585  0.30128676
#> 7   007  1.460962332  2.54720967
#> 8   008  1.081189593  0.42161656
#> 9   009  0.578154456 -0.71674569
#> 10  010  1.712342794  0.35081463
#> 11  011 -0.869997634  1.56818247
#> 12  012  0.604070396 -0.02672035
#> 13  013 -0.757146433 -0.38221618
#> 14  014 -0.244199632 -0.04063259
#> 15  015 -0.176959224  0.32420574
#> 16  016  0.117419473 -0.76042573
#> 17  017  1.986032061  0.59263097
#> 18  018  0.561277093  0.18013256
#> 19  019  0.782557479  0.07865119
#> 20  020  0.744753018  1.15911573
#> 21  021  1.540022324 -0.50442369
#> 22  022 -0.522901050 -1.63594022
#> 23  023 -0.404372237 -0.03585991
#> 24  024  0.765975972 -0.81716595
#> 25  025  0.656103139  0.77582852
#> 26  026  0.861618590 -0.62328899
#> 27  027 -0.848189969 -0.86483402
#> 28  028 -1.929619759 -1.86364374
#> 29  029  1.767688248  1.26743003
#> 30  030  0.448962583  1.28384051
#> 31  031 -0.924529534 -1.19331792
#> 32  032  0.884999912  0.96366732
#> 33  033  0.837293333  0.78048222
#> 34  034  0.313591211  0.96228941
#> 35  035  0.343081668  0.79527326
#> 36  036  0.453368566  0.97656782
#> 37  037  0.828444625  0.86613516
#> 38  038 -0.006132247 -0.77290441
#> 39  039  0.068204533  0.22637052
#> 40  040 -0.137496867 -0.49753399
#> 41  041 -0.507541226 -0.59051493
#> 42  042 -0.733625338 -0.33529345
#> 43  043  0.335273196  1.26186094
#> 44  044  2.057446487  0.63163620
#> 45  045  1.736741805  1.22450438
#> 46  046 -1.141760718 -1.69158646
#> 47  047  0.986942203  0.34815295
#> 48  048  0.602418013  0.87921322
#> 49  049 -0.110724247  0.39306518
#> 50  050  0.394891163  0.19628423
#> 51  051  0.407349583  1.04459995
#> 52  052 -0.404792246 -0.07824708
#> 53  053  0.448212786  0.12959065
#> 54  054  0.469505269 -0.24532287
#> 55  055 -0.038214908  0.14048337
#> 56  056 -1.741121824 -0.95923665
#> 57  057  1.840268170  0.04188894
#> 58  058 -0.251615515  0.37094383
#> 59  059 -0.638674754  0.10290934
#> 60  060 -1.153336164 -0.40973724
#> 61  061 -0.254149678  0.49149775
#> 62  062 -0.632765588  0.43067027
#> 63  063  0.201049632 -0.39685507
#> 64  064 -0.457582207 -0.53529225
#> 65  065  0.362503384  0.15254924
#> 66  066  0.338648309  1.06085079
#> 67  067  0.133780532 -0.08831835
#> 68  068  0.779602095  0.74388430
#> 69  069 -2.121236050 -2.01405114
#> 70  070  0.356796653 -0.78042224
#> 71  071 -1.057984562  0.10569629
#> 72  072  0.624864065 -1.07490174
#> 73  073 -2.374710716 -1.05858977
#> 74  074 -0.310434203 -0.33929532
#> 75  075  0.876796358  1.90431675
#> 76  076  1.654423208 -0.51559604
#> 77  077 -0.860745497 -1.97038320
#> 78  078  1.132021385 -0.37358985
#> 79  079  0.554199693  0.08486162
#> 80  080 -0.004158898 -0.35611777
#> 81  081 -0.539806096  0.24137491
#> 82  082 -0.842652570 -1.24772646
#> 83  083 -0.096827879  1.44262046
#> 84  084  0.567119009  2.57883264
#> 85  085 -1.308956267  1.01786884
#> 86  086  1.219332776 -0.29066858
#> 87  087  2.406323130  0.56188111
#> 88  088  1.089836153  1.57189494
#> 89  089  0.984090599  0.65483418
#> 90  090  1.484871228  0.35638651
#> 91  091 -2.715442247 -1.66537779
#> 92  092 -0.846507395 -1.43564917
#> 93  093  1.357021954  0.74642937
#> 94  094 -0.359898239 -0.25659142
#> 95  095  1.967916857  0.82219403
#> 96  096  1.260259188  0.59452690
#> 97  097 -0.014991059  0.87274189
#> 98  098 -0.266926733 -1.30904268
#> 99  099  0.575138994 -0.06066056
#> 100 100 -0.097177587  0.89011465