Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5333188
#> Y_2 0.5333188 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>         Y_1        Y_2
#> 1 1.6787643  1.6830980
#> 2 0.4187897 -1.9144169
#> 3 0.6776575  0.1128619
#> 4 0.5510104 -1.2004926
#> 5 0.6696436  1.3406057

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.079792642  0.40960116
#> 2   002 -2.431145998 -2.50910235
#> 3   003  0.051419072  1.20142096
#> 4   004  1.227672303  0.43114365
#> 5   005  0.510652651  1.14510514
#> 6   006 -0.029290234  0.02868453
#> 7   007  0.197242581  0.36379545
#> 8   008  0.942172950  0.02510385
#> 9   009 -0.877080241 -0.97392159
#> 10  010  0.898644004  0.30878973
#> 11  011  1.991440617  0.51643582
#> 12  012  0.475752686 -0.24088046
#> 13  013  0.875815804  0.67103398
#> 14  014 -0.709464340 -0.12774577
#> 15  015  0.071842552  0.75351372
#> 16  016  0.452367615  0.38171437
#> 17  017 -0.016489977 -0.27121560
#> 18  018  1.065911380 -0.33233188
#> 19  019  0.174180426  1.59541438
#> 20  020 -0.675947931 -0.35917518
#> 21  021  0.005012577 -0.30284481
#> 22  022  0.230992688 -0.05070656
#> 23  023  0.330428306 -0.31158070
#> 24  024 -0.640260287 -0.30555329
#> 25  025 -0.202999065 -0.67316440
#> 26  026  1.142923831  0.61302877
#> 27  027  0.246646605  1.96477631
#> 28  028 -1.408646024 -0.19646472
#> 29  029 -0.626733485  0.43507825
#> 30  030 -1.000313794  0.06060587
#> 31  031  0.457433898  0.80041701
#> 32  032  0.965343847  1.32122222
#> 33  033 -0.282231076  1.06300278
#> 34  034 -0.587816245  0.62559797
#> 35  035  0.622114859  0.70846180
#> 36  036  0.449798968  0.28875584
#> 37  037  0.064702640  1.45397341
#> 38  038 -1.086743309 -0.54110703
#> 39  039  0.598484975  0.27565806
#> 40  040 -0.477189839  0.38999908
#> 41  041  1.069469139 -0.84906603
#> 42  042  0.521447702  1.53040036
#> 43  043 -1.709099828  0.52308954
#> 44  044  0.342226462  0.33815733
#> 45  045  0.602247253 -0.60410325
#> 46  046  0.789076487 -0.21778797
#> 47  047  1.071544539 -0.66682981
#> 48  048 -0.852292278 -0.48549539
#> 49  049 -1.325576773 -0.51492265
#> 50  050  0.266188783  1.58003832
#> 51  051 -0.476220891 -1.18373117
#> 52  052 -1.937817781  0.78880426
#> 53  053  0.526848766 -0.24727105
#> 54  054 -0.228607212 -1.15050204
#> 55  055 -0.083051941  0.98622743
#> 56  056 -1.349521230 -2.98398876
#> 57  057 -0.350813667  0.24835934
#> 58  058 -1.642288653 -1.24906366
#> 59  059 -1.308387411  0.01321079
#> 60  060 -0.188193007 -0.23848291
#> 61  061 -1.620358070 -2.54420158
#> 62  062 -1.268319706 -1.42724574
#> 63  063  0.912513560  0.03868397
#> 64  064  1.000253315  1.37513637
#> 65  065 -1.973337570  0.18633250
#> 66  066  0.150936741 -0.46238243
#> 67  067 -1.442503846 -1.58945261
#> 68  068 -1.197326369  0.01015780
#> 69  069  0.636539033 -0.72163325
#> 70  070 -1.570699970 -2.42827158
#> 71  071 -0.441228216 -1.09756566
#> 72  072 -1.027313748  0.38703617
#> 73  073  0.595435397  1.42342481
#> 74  074  0.640657666  2.35543367
#> 75  075 -0.637657448  0.07591696
#> 76  076  0.670352065 -0.47174978
#> 77  077  1.049370630  0.58565343
#> 78  078  0.548980380  0.61717098
#> 79  079 -0.685532775 -0.73334842
#> 80  080  0.436525565  1.02093145
#> 81  081  0.422550116 -0.05906908
#> 82  082 -1.385524709 -1.04614707
#> 83  083 -1.276942942  0.01036760
#> 84  084 -0.526099280  0.30418240
#> 85  085  0.946439195 -1.00292866
#> 86  086  0.097202209  0.42492137
#> 87  087  1.481232976 -0.32678500
#> 88  088  1.701932078  0.81228481
#> 89  089 -2.049498167 -2.61738415
#> 90  090 -0.456388034 -1.79748191
#> 91  091  0.715410526  1.02354522
#> 92  092 -1.270091594 -2.40086176
#> 93  093 -0.214131228 -0.99489970
#> 94  094  0.586652920 -0.07484407
#> 95  095  1.035368718  0.56045624
#> 96  096  0.934696102  1.17480611
#> 97  097 -1.012740230 -0.42986655
#> 98  098 -0.587177582 -0.35646682
#> 99  099  1.372566974 -0.17295842
#> 100 100 -0.489463814  0.39659303

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.60650906 -1.609844506
#> 2   002  1.31952732  0.322497835
#> 3   003 -0.04188495  1.715118850
#> 4   004 -0.37738681  0.747066520
#> 5   005  0.40475423 -0.437710582
#> 6   006  1.34290334 -0.621881790
#> 7   007 -0.08261866  0.520453076
#> 8   008  0.33435412  0.817108305
#> 9   009 -1.09577986 -0.278063947
#> 10  010 -0.56835019  0.406669047
#> 11  011  0.09943725 -0.212791790
#> 12  012 -0.09978274 -0.721151625
#> 13  013 -0.07134971 -0.500752515
#> 14  014  0.36195885 -0.382058096
#> 15  015  1.56696106  1.810863578
#> 16  016  1.10298347  0.699123313
#> 17  017 -0.30685438 -0.931816667
#> 18  018  0.55282066  0.009852176
#> 19  019 -2.17065752 -0.278634095
#> 20  020 -0.91284954 -0.041628346
#> 21  021  1.67017595  0.956213998
#> 22  022  0.13679455  0.118374990
#> 23  023 -0.54524511  0.764607858
#> 24  024  1.17857837  0.948964237
#> 25  025 -0.99873188 -0.102113941
#> 26  026  0.93028919 -0.431773622
#> 27  027 -1.22328126 -0.973345219
#> 28  028  0.55323919  1.182642807
#> 29  029  1.16950178  0.579851234
#> 30  030  0.37585576  0.214302247
#> 31  031  0.46517094  0.563186821
#> 32  032  1.83169881  1.001389123
#> 33  033 -0.80065517 -0.089730419
#> 34  034  0.95462137  0.596233975
#> 35  035 -0.82957123 -0.348631905
#> 36  036 -0.44977622 -0.556892354
#> 37  037 -0.31542284 -1.029183925
#> 38  038  0.13676365  1.172935092
#> 39  039  1.18065349  0.223704513
#> 40  040  2.09963326  1.288453147
#> 41  041 -1.10248723  0.036119922
#> 42  042 -0.82893334 -0.347364733
#> 43  043  1.65493677  0.589955992
#> 44  044  1.20427836 -0.059540799
#> 45  045 -0.47066115 -1.658421934
#> 46  046  0.53129671 -1.023928868
#> 47  047 -0.62958291 -0.431669145
#> 48  048  0.10478598 -0.412093113
#> 49  049  0.64536567  1.283787013
#> 50  050  0.70447591  1.216003158
#> 51  051 -1.37483619 -2.254052504
#> 52  052 -0.24751134 -0.190221019
#> 53  053  1.19049014  0.870509396
#> 54  054  0.92658052  0.702649255
#> 55  055  0.47491904  1.232406180
#> 56  056  0.82020818  0.449166994
#> 57  057  1.08588411 -0.988339763
#> 58  058  0.42353365 -0.275381755
#> 59  059 -2.83224159 -2.994788003
#> 60  060  0.76148768  0.620513806
#> 61  061  0.17512866  0.993651769
#> 62  062 -0.34396621 -0.797631646
#> 63  063 -2.23090133 -1.141112234
#> 64  064 -0.71648155 -0.599487827
#> 65  065  0.20626034  0.499063674
#> 66  066 -0.99583445 -0.355216688
#> 67  067 -0.28015388  0.749003900
#> 68  068  1.25752319  0.079548267
#> 69  069 -0.14698506  0.319450565
#> 70  070 -1.04550993 -1.372905338
#> 71  071  1.88030434  0.710362488
#> 72  072 -0.54603274 -0.591161553
#> 73  073 -1.03110609 -1.767226355
#> 74  074 -0.58170250  0.797497395
#> 75  075  0.36025797 -0.552600300
#> 76  076  0.67816070  1.143080321
#> 77  077  1.01172874  1.799122454
#> 78  078 -0.86691516 -1.591305119
#> 79  079  1.45345488  0.705439562
#> 80  080 -1.05557117 -1.781527545
#> 81  081  0.58046568  0.558911962
#> 82  082 -2.13282840 -0.877373058
#> 83  083  0.94096474  2.290682284
#> 84  084  1.27843398  0.327189396
#> 85  085  1.33341288 -0.605187911
#> 86  086 -0.15492031  1.110379596
#> 87  087 -0.11657649  1.127433352
#> 88  088  0.11139881 -0.058565002
#> 89  089 -0.87455711 -0.857801741
#> 90  090 -0.12009183 -1.053549793
#> 91  091  1.07111343  0.129095537
#> 92  092 -2.69663265 -1.708664472
#> 93  093  1.79368620  0.926405990
#> 94  094 -0.62457710 -0.155028335
#> 95  095 -0.16228780  0.098004377
#> 96  096 -0.05087344  0.068211356
#> 97  097  0.51635813  1.155802308
#> 98  098 -1.05659664  1.353282740
#> 99  099  0.15476235  0.426194898
#> 100 100  1.34315522  0.595191460