Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5469114
#> Y_2 0.5469114 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.3874612  0.7579962
#> 2  0.7662682  0.6468556
#> 3 -0.2113319 -0.8343189
#> 4  1.0396786 -0.3814510
#> 5 -0.7416959  0.4102361

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.06705099  0.70920421
#> 2   002 -2.40866809 -0.54921990
#> 3   003 -1.27920263  0.41280268
#> 4   004  0.36887557  0.95552701
#> 5   005  0.39205738 -1.04704469
#> 6   006 -0.17304101 -0.72017591
#> 7   007  0.18008664 -0.03890887
#> 8   008  0.79863630  0.57436394
#> 9   009  1.49102974 -0.71574229
#> 10  010  2.36119414  1.02917688
#> 11  011 -0.79356918 -0.62598068
#> 12  012 -0.59210855 -1.05653829
#> 13  013  0.04485207 -0.13135956
#> 14  014  1.19973143  1.52074518
#> 15  015  1.13365309  1.24838149
#> 16  016  0.07548994 -0.59389362
#> 17  017 -0.06448430  0.03242696
#> 18  018  1.07857830  1.14747461
#> 19  019  0.83350963 -1.68272719
#> 20  020 -0.64175013 -0.44935240
#> 21  021 -0.66488896 -0.58180319
#> 22  022  0.14746428  2.22129249
#> 23  023 -0.69281262 -0.66772504
#> 24  024  0.23782335  0.57395279
#> 25  025 -1.14037211 -0.17411108
#> 26  026 -0.83218928 -0.28872501
#> 27  027 -0.80673920 -1.19900207
#> 28  028 -0.84515344  0.38546122
#> 29  029  0.38184616 -0.40402172
#> 30  030 -0.48317940  0.34830294
#> 31  031  0.83599187  0.40664245
#> 32  032  0.05529756  0.48568126
#> 33  033 -2.16906339 -1.00875440
#> 34  034  1.30231482 -0.34108875
#> 35  035  1.11273829  0.39942057
#> 36  036  1.41425490 -0.12283649
#> 37  037 -2.25110725  0.11334841
#> 38  038  0.22569576  0.11354539
#> 39  039  1.48529896 -0.08904963
#> 40  040  1.67296039  1.63981431
#> 41  041 -0.67400262 -0.14118091
#> 42  042 -0.85516297 -1.87753571
#> 43  043 -0.18834622 -1.40870119
#> 44  044 -0.21801231  0.70039598
#> 45  045  0.88523008 -0.15402192
#> 46  046 -0.61575547 -2.21215272
#> 47  047  0.40436753 -1.86007044
#> 48  048 -0.07919691 -0.45360661
#> 49  049 -1.75842583 -0.82258623
#> 50  050 -0.53525269 -0.76241379
#> 51  051 -0.83258203 -0.30201232
#> 52  052  0.66126422 -0.29196794
#> 53  053 -0.91474326  0.02016099
#> 54  054 -0.90047839 -2.09529375
#> 55  055 -0.70601170  0.87567544
#> 56  056 -1.60629429 -1.32564595
#> 57  057  0.07372530 -0.51940840
#> 58  058  0.39138793 -0.61685703
#> 59  059  1.26506428  0.90037344
#> 60  060 -1.27764643 -0.42988348
#> 61  061  0.48005863  1.24306431
#> 62  062 -0.67869414 -1.06033390
#> 63  063  0.80690537  0.12239137
#> 64  064 -1.55295828 -0.85228406
#> 65  065  0.36309119  2.45189312
#> 66  066  1.10071596  1.50953087
#> 67  067 -0.36382413  0.54153574
#> 68  068  0.76624734  1.25512955
#> 69  069  0.27508985 -0.44161541
#> 70  070 -0.43326472 -0.22707542
#> 71  071 -0.58694997 -1.22129331
#> 72  072 -0.78866515  0.40937347
#> 73  073 -1.24088581  0.12234658
#> 74  074  0.66607104 -0.73975955
#> 75  075  0.63004576  0.97756591
#> 76  076 -0.94088667 -0.24975042
#> 77  077  0.82590593 -0.84232424
#> 78  078 -0.75713260 -1.75848458
#> 79  079  1.00854794  0.47730070
#> 80  080  0.42510470 -0.59916393
#> 81  081 -1.21535389  0.28721016
#> 82  082 -1.44541487 -0.12461317
#> 83  083  1.39143631  1.04452281
#> 84  084 -1.33203353 -0.88639263
#> 85  085  0.19920241  1.42461952
#> 86  086  1.39457503  1.69546713
#> 87  087  0.08764192  0.42213116
#> 88  088 -0.15467741 -0.36681151
#> 89  089  0.63974249  1.49327710
#> 90  090  0.10517157  0.75999579
#> 91  091  1.04598284  0.38544370
#> 92  092 -1.49986956 -0.82170645
#> 93  093  0.75982423  0.29587386
#> 94  094 -1.04825613  0.15304960
#> 95  095  0.04720935 -0.88656324
#> 96  096  1.58618850  1.90222994
#> 97  097  1.21886994  1.07464160
#> 98  098 -1.44839223  0.43512424
#> 99  099 -0.35327638  0.23134275
#> 100 100  0.43560115  0.07620282

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.888547886 -0.76201103
#> 2   002  0.104304362 -0.79628151
#> 3   003 -1.015148052 -0.41399067
#> 4   004  1.028811996  1.67182932
#> 5   005  0.849690794  0.40650642
#> 6   006 -0.237989039 -0.20911984
#> 7   007  0.739037461  0.18966127
#> 8   008  0.067992815 -0.96678914
#> 9   009  0.036325788  0.56743167
#> 10  010 -0.476960585 -0.53336991
#> 11  011  0.195638193 -0.06853136
#> 12  012  1.890097322  0.24967672
#> 13  013 -1.194880548  2.65547922
#> 14  014  1.355750801  1.87680695
#> 15  015  2.363229301  3.71150788
#> 16  016  2.493065680  1.39838667
#> 17  017 -0.885708664  0.45752350
#> 18  018  0.346990754  1.13949414
#> 19  019  0.210263726 -0.03479668
#> 20  020  1.280976718 -0.47602327
#> 21  021  1.089755530 -0.80492550
#> 22  022 -0.916818218 -0.35580476
#> 23  023  0.553799832 -0.69598700
#> 24  024  0.198633508  1.06075730
#> 25  025  1.777338434  2.59316841
#> 26  026  2.844244126  0.73631318
#> 27  027 -2.400958916 -0.27907044
#> 28  028 -0.565438815 -0.61771155
#> 29  029 -2.081051036 -1.04132638
#> 30  030  1.421083368  0.51628425
#> 31  031  1.270053126 -0.45046896
#> 32  032  0.257464443 -0.64194180
#> 33  033 -1.817551578 -0.95192549
#> 34  034  0.276580267  1.43338794
#> 35  035  0.216175528  0.61530632
#> 36  036 -1.113359380  0.05697442
#> 37  037 -0.545797958 -1.53521877
#> 38  038 -0.911597859  0.03584942
#> 39  039  0.148108186 -0.31135636
#> 40  040 -0.679722541 -0.63478455
#> 41  041 -1.203494318 -0.75501054
#> 42  042  1.143829120  0.87787364
#> 43  043 -0.923309857 -0.39083501
#> 44  044  1.568320362  1.03972490
#> 45  045 -0.516917604  0.92535662
#> 46  046 -0.766452445 -2.11047312
#> 47  047  2.739180682  1.58316815
#> 48  048 -0.728022777 -0.95341225
#> 49  049  0.515259326  0.56030443
#> 50  050 -0.570609172  0.05958236
#> 51  051 -1.614835849 -0.92264424
#> 52  052 -0.681261302 -2.08610731
#> 53  053  0.115697821  0.85051248
#> 54  054 -0.160080996  0.25521908
#> 55  055  0.158640188 -0.62001604
#> 56  056  1.443748865  1.60678452
#> 57  057 -1.572929169 -0.20373110
#> 58  058 -0.430866543  0.05774021
#> 59  059  1.099986780  1.11999976
#> 60  060 -1.526747846  1.85477967
#> 61  061  1.736312721  2.28887786
#> 62  062 -1.535660056 -0.88883231
#> 63  063 -1.802298891 -1.96531740
#> 64  064  1.420517505  1.83070849
#> 65  065 -0.121466267 -1.51603695
#> 66  066  0.529208676  1.89043565
#> 67  067  0.441916383  0.03899671
#> 68  068 -0.900684497 -1.01227448
#> 69  069  0.930477652 -1.20792676
#> 70  070 -0.007122554 -0.31943332
#> 71  071  1.521850484  1.16040204
#> 72  072  0.377190917  0.07857531
#> 73  073  1.551139840  1.52740094
#> 74  074 -0.235859502 -0.98344324
#> 75  075 -0.261498643 -0.17761012
#> 76  076 -1.352891958  0.50502544
#> 77  077 -0.753183071 -0.59083768
#> 78  078 -0.577695597 -2.83301966
#> 79  079  0.101257410 -1.00598470
#> 80  080 -1.612093066 -1.18340570
#> 81  081 -0.558026022 -0.75743583
#> 82  082  1.772717326  1.48742645
#> 83  083  0.105928344  0.26561638
#> 84  084 -0.643628018  1.54935926
#> 85  085 -0.041575961  0.98186666
#> 86  086  0.077750065  2.29433909
#> 87  087  1.465737704 -1.29577525
#> 88  088  0.887092208  0.65444988
#> 89  089 -1.380155174 -1.25129486
#> 90  090 -0.371822614  0.17785569
#> 91  091 -1.188879270  1.03783016
#> 92  092 -0.902119867  2.57817201
#> 93  093  0.277079306 -0.61096648
#> 94  094  1.675439391  1.51492780
#> 95  095 -0.930951192 -1.05924343
#> 96  096  0.461370594 -0.09138850
#> 97  097  1.149282573  0.54610351
#> 98  098  1.100636355 -0.27324845
#> 99  099  1.491661167 -0.80569841
#> 100 100 -0.154547286 -0.71045414