Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5148574
#> Y_2 0.5148574 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.1553602 -0.87333078
#> 2 -0.3738710 -0.48419616
#> 3  0.9242197  0.67906450
#> 4 -2.3410584 -1.21603167
#> 5 -0.5330600  0.04024834

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.91040554 -0.467785632
#> 2   002 -1.48966897 -2.000079480
#> 3   003 -1.21978738  0.267399881
#> 4   004  1.45096064  0.292283327
#> 5   005 -0.80238151 -2.733305093
#> 6   006 -1.79649497 -2.273695256
#> 7   007 -0.26928696  1.117578385
#> 8   008 -0.72390035 -0.205033894
#> 9   009  1.16746672  0.077371963
#> 10  010  2.19193614  1.158646724
#> 11  011  1.22368209  0.635034319
#> 12  012  0.33260801  1.226847312
#> 13  013 -1.70962212 -0.139181732
#> 14  014 -0.62560208 -1.917209300
#> 15  015  0.24062300  2.386938693
#> 16  016 -1.27586776 -0.399689904
#> 17  017 -0.47653369  0.398185583
#> 18  018 -0.13458001 -1.325938771
#> 19  019  0.43261083  1.218372882
#> 20  020 -0.32933079 -1.638198409
#> 21  021  1.65343259  1.285649418
#> 22  022 -2.11741303 -0.488440982
#> 23  023 -0.17784081 -0.158644544
#> 24  024  1.43846329  0.763610647
#> 25  025  1.55404680  1.842635728
#> 26  026 -0.07690001 -0.086925966
#> 27  027  1.40847256  1.441517029
#> 28  028 -0.90396903 -1.160596344
#> 29  029 -0.67821352  0.087848883
#> 30  030  0.66785113  0.769118663
#> 31  031 -1.60128182 -1.646861541
#> 32  032  1.18134689  0.383972859
#> 33  033  0.37853279 -0.750183446
#> 34  034  0.24899286  0.500247973
#> 35  035  0.68537264  0.243076182
#> 36  036  0.87483241  0.196059785
#> 37  037  1.77232505  1.592898429
#> 38  038 -0.25364853 -1.820572176
#> 39  039  0.25089565 -1.161426785
#> 40  040 -0.21564086 -0.908801761
#> 41  041  0.05341970 -0.503890219
#> 42  042 -0.17173686  0.175619572
#> 43  043  0.15717794 -0.753029700
#> 44  044  1.75274532  2.895755332
#> 45  045  3.55503510  0.604970463
#> 46  046  0.27365047 -0.227474687
#> 47  047 -1.59358528 -0.378750112
#> 48  048  0.81979573  0.249125705
#> 49  049 -0.61455926 -0.400436033
#> 50  050  0.45273382  0.164348199
#> 51  051 -0.24024142 -1.463321264
#> 52  052  0.93283442  1.535972356
#> 53  053  2.41142491  1.518007256
#> 54  054  0.44691402  0.878384137
#> 55  055  0.43932428  0.105631271
#> 56  056  0.53055286 -0.390483341
#> 57  057 -0.25830296  0.199751950
#> 58  058 -0.93630368 -0.930947947
#> 59  059  0.42820001  0.637570954
#> 60  060 -3.19356139 -3.392182539
#> 61  061 -0.03522039 -0.961156754
#> 62  062  0.27010314  0.969253786
#> 63  063  0.30702971  0.043686469
#> 64  064  0.81327547  0.271472488
#> 65  065  1.70381889 -0.003450782
#> 66  066  1.20713692  1.091945353
#> 67  067  0.22805722  0.586175214
#> 68  068 -0.29672894  0.288293093
#> 69  069 -0.19359207  1.752540307
#> 70  070  1.04604990 -0.356855936
#> 71  071  0.76261133 -0.558320033
#> 72  072  1.49905889 -0.071444104
#> 73  073  0.54198795  0.535559235
#> 74  074 -1.14945830 -0.259605242
#> 75  075 -0.55528502 -1.355835169
#> 76  076 -1.10854803 -0.853558741
#> 77  077  1.09506637  2.883300134
#> 78  078 -2.59554275 -0.808308872
#> 79  079  0.73240675  1.141068289
#> 80  080 -0.05835498  0.553561630
#> 81  081  1.41856217  1.967084867
#> 82  082 -2.28730099 -0.515730854
#> 83  083 -2.04654151 -1.452306848
#> 84  084 -0.33971307 -0.377516591
#> 85  085 -0.69939719 -0.372940690
#> 86  086  1.56860224  0.432707457
#> 87  087 -0.68327956  0.154899808
#> 88  088 -0.49754197 -0.946165258
#> 89  089 -0.21922644  0.284215237
#> 90  090 -0.28711317  0.720692368
#> 91  091 -1.02929265  0.089826741
#> 92  092  1.49800971  0.244981930
#> 93  093  2.00797593  1.617163472
#> 94  094  0.64853206 -0.167797045
#> 95  095 -0.60340228 -0.642233893
#> 96  096  1.39102944  0.135773488
#> 97  097  1.61891828  0.366475569
#> 98  098  0.10102633 -0.903017717
#> 99  099 -1.29093558 -1.860802446
#> 100 100  0.19244422  0.036355481

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  1.143711393  1.30301387
#> 2   002  2.335297669 -0.03850944
#> 3   003 -0.954515234  1.34611938
#> 4   004 -0.133074497 -0.47893897
#> 5   005  1.703090332  0.93508856
#> 6   006 -1.110030471 -1.45923344
#> 7   007 -0.601517265  0.38469441
#> 8   008  1.576598235  0.98082291
#> 9   009  1.889604646  1.40577524
#> 10  010 -0.202459742  0.19956820
#> 11  011 -0.048451063 -0.58181238
#> 12  012 -0.419244729 -0.60787645
#> 13  013 -0.987140799 -1.13619131
#> 14  014  0.163835269 -0.60888996
#> 15  015 -0.629806497 -1.56103760
#> 16  016  0.755894911  0.31876477
#> 17  017 -1.826246411 -0.52894705
#> 18  018  2.101434683  2.69239534
#> 19  019 -0.543421892 -1.44118781
#> 20  020 -1.014035212 -1.17319258
#> 21  021 -0.570729409  0.26818039
#> 22  022  0.505643695 -0.28112336
#> 23  023 -0.563266278 -0.37922060
#> 24  024  0.816010064  0.68438964
#> 25  025  1.558413334  1.08347545
#> 26  026  0.145381768  1.00323219
#> 27  027 -0.087957114 -0.31620967
#> 28  028  0.037508067  2.37783872
#> 29  029  0.667579366  1.09416289
#> 30  030 -0.154486719  1.12894412
#> 31  031  0.395715400 -0.12330600
#> 32  032 -2.158545981 -1.61466589
#> 33  033 -0.548458723 -0.31706676
#> 34  034 -0.596682836 -0.37710041
#> 35  035 -0.077891219  1.25948829
#> 36  036 -0.659618281  1.43091568
#> 37  037 -0.432167532  0.72493652
#> 38  038 -0.477030295  0.26588640
#> 39  039  0.052616012 -1.10414501
#> 40  040  1.114747318  1.07533286
#> 41  041  1.630200082  0.64556137
#> 42  042  0.340917791  0.28690731
#> 43  043 -0.955849575 -0.56640758
#> 44  044  0.733145897 -0.30064525
#> 45  045  1.359534956 -0.63000657
#> 46  046 -0.936949386  0.91473468
#> 47  047  1.749779951  1.36601308
#> 48  048 -0.192855418 -1.17697048
#> 49  049  1.160154474  0.01103885
#> 50  050  2.195770955  1.41395604
#> 51  051  0.082353704 -2.02106849
#> 52  052 -1.011684139 -1.26387476
#> 53  053 -1.349603078 -0.45839072
#> 54  054  0.130263408  0.54704376
#> 55  055 -0.066487762  1.18617984
#> 56  056  0.766058317 -0.24055342
#> 57  057 -0.189667615  0.54841344
#> 58  058 -1.202336179 -0.03168128
#> 59  059 -0.960528812 -1.19713489
#> 60  060 -0.135341162  1.76861683
#> 61  061 -1.118264851 -0.62850728
#> 62  062  0.694254820 -0.01274268
#> 63  063 -0.289225662  0.12020052
#> 64  064 -0.346821809  0.32629516
#> 65  065  0.893014630  0.90303050
#> 66  066 -0.356835954 -0.57676547
#> 67  067 -0.564207275  1.19122764
#> 68  068  1.951383917  0.75180262
#> 69  069 -0.407571299 -0.19012316
#> 70  070 -0.011009182 -1.65445305
#> 71  071  1.499921328  0.96654469
#> 72  072  0.160993606 -0.49319931
#> 73  073  0.278087510  1.44586462
#> 74  074 -2.428332309  0.24408079
#> 75  075 -0.846214246 -1.63707738
#> 76  076 -0.056321115 -1.19921636
#> 77  077  0.088510666  0.34700758
#> 78  078  0.033715196 -0.59263520
#> 79  079  0.863204632  1.30840960
#> 80  080  0.166616093  0.30355554
#> 81  081 -0.069857636 -0.75850735
#> 82  082  0.663661489  1.13397307
#> 83  083  1.639623037  0.02492506
#> 84  084  0.484554732  0.10362601
#> 85  085 -0.782369089 -0.78964854
#> 86  086  0.160370496 -0.03999940
#> 87  087 -0.007117787  0.19013791
#> 88  088  1.096511791  1.61227287
#> 89  089  0.184602239  0.61278341
#> 90  090  1.325880480  1.10143623
#> 91  091  1.926978237  2.04119819
#> 92  092 -0.599318119 -0.66034098
#> 93  093  2.193918593 -0.42268680
#> 94  094 -0.449814394 -0.26784075
#> 95  095 -2.235219027 -1.38950107
#> 96  096 -1.352662754 -1.23883979
#> 97  097  0.833832328  0.69034211
#> 98  098  0.455753794  1.16154476
#> 99  099 -1.427553030 -0.74934673
#> 100 100 -0.208535354 -1.22441270