Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5442138
#> Y_2 0.5442138 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.3566323  1.47677569
#> 2 -0.0727175 -0.03769493
#> 3  1.2008626 -0.25614485
#> 4  1.8117885  0.97977783
#> 5  1.1605143  0.23714486

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.478219988  0.96085030
#> 2   002  0.313276110 -0.22414355
#> 3   003 -1.975544402 -1.77018895
#> 4   004 -0.146901214  0.09724039
#> 5   005 -0.531569981  1.20867606
#> 6   006  0.241255863 -0.24181026
#> 7   007 -0.721431076 -0.55833557
#> 8   008 -1.291897797 -0.76491533
#> 9   009  0.819446059  0.12210551
#> 10  010 -0.737263775 -1.61802749
#> 11  011  1.558579802 -0.12028197
#> 12  012  0.243226910 -0.71252843
#> 13  013 -0.328942579 -0.13964499
#> 14  014  1.495147849  0.45434817
#> 15  015  2.076529473  1.55490286
#> 16  016 -0.466480561  0.29986717
#> 17  017 -0.636693751  0.20631179
#> 18  018  0.540671661  0.73626224
#> 19  019  0.257577055  0.55785108
#> 20  020  0.368643601  0.92484187
#> 21  021 -1.372050746 -1.79963732
#> 22  022  0.528855025 -0.82653113
#> 23  023 -0.631736743  1.60463091
#> 24  024  0.043221521 -0.43927849
#> 25  025 -0.197152087 -0.38751074
#> 26  026  1.272930728  0.41403503
#> 27  027  1.619796979  1.92066021
#> 28  028 -1.204118056 -0.49404849
#> 29  029 -0.207138172 -1.48029894
#> 30  030 -1.156450967 -0.76646583
#> 31  031  1.216092009  0.95789623
#> 32  032 -0.980175527 -1.14426733
#> 33  033 -2.386063456 -0.58195056
#> 34  034 -0.837814304  0.13724003
#> 35  035 -0.808215127 -1.11755615
#> 36  036 -1.192150491 -1.03184698
#> 37  037 -1.488557789 -0.69835441
#> 38  038  2.603211954  1.28320251
#> 39  039  0.462523072  1.13224828
#> 40  040  0.795098936 -0.31990308
#> 41  041 -1.135953675 -2.95388407
#> 42  042 -0.141742562 -0.75309492
#> 43  043  0.440087798  0.77320572
#> 44  044 -1.576949299 -1.14617888
#> 45  045 -0.662986298 -0.88123699
#> 46  046  0.944460159  1.66421618
#> 47  047  1.467606219  0.39474740
#> 48  048 -1.650041433 -2.87313163
#> 49  049  0.843588099  1.00692700
#> 50  050  0.689235089 -0.32813925
#> 51  051 -0.027381213 -0.28712928
#> 52  052  1.308497029 -0.39134653
#> 53  053 -0.951351033  2.07621052
#> 54  054  0.842988021  0.44538816
#> 55  055  0.117376861 -0.33337575
#> 56  056  0.693097524  0.11846595
#> 57  057  0.988103528  2.16195150
#> 58  058  0.820405451  1.54674051
#> 59  059 -0.470869117  0.28161485
#> 60  060 -1.329061712  0.20042233
#> 61  061  0.059530417  0.18904665
#> 62  062  0.113179112 -0.59122409
#> 63  063  2.363579118  1.09317280
#> 64  064 -2.085969064  0.16988559
#> 65  065 -0.516331447  0.08188429
#> 66  066 -0.189747795 -0.36973222
#> 67  067  1.235903948  0.75630958
#> 68  068  0.135276229  1.92995936
#> 69  069  0.476306725 -0.01724792
#> 70  070 -0.097971075 -1.89475393
#> 71  071  1.431178486  0.53645363
#> 72  072  1.189382457  2.22374474
#> 73  073 -1.530831886 -2.06941088
#> 74  074 -1.207168205 -0.67363481
#> 75  075 -1.363182905 -0.68325399
#> 76  076 -0.167215883  0.76549380
#> 77  077  2.123638869  0.17466273
#> 78  078 -2.883218883 -0.90668730
#> 79  079 -2.437683751 -1.74393382
#> 80  080  0.555899467  1.34577337
#> 81  081 -1.337281686 -0.97640049
#> 82  082 -0.039455458 -0.45435853
#> 83  083 -1.211360227 -1.03087953
#> 84  084  1.370392828  0.86995775
#> 85  085 -0.102752019 -0.80841555
#> 86  086 -0.951388790 -0.70933848
#> 87  087 -0.605559615  0.72239254
#> 88  088  0.434494952  0.27455337
#> 89  089  0.300140266  0.15227358
#> 90  090 -0.943461655 -2.31815721
#> 91  091 -1.045118912 -0.27525773
#> 92  092  0.131981693  0.26065940
#> 93  093 -1.265216654 -1.01746471
#> 94  094 -1.249096743 -1.78932987
#> 95  095  0.478089948  0.60823580
#> 96  096 -0.258132048 -0.24187049
#> 97  097  0.001293626 -0.65155068
#> 98  098  0.242233088  0.22470922
#> 99  099  1.919324237  2.14322560
#> 100 100  0.131033328  0.16044485

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.101156049 -1.50305552
#> 2   002 -0.407654919  0.11090727
#> 3   003 -0.167461347  0.71075066
#> 4   004  0.656031494 -0.35511844
#> 5   005  1.140762926  1.66578863
#> 6   006  0.721248431 -0.73188763
#> 7   007 -0.721025353 -1.21326468
#> 8   008 -0.978325357 -0.57226034
#> 9   009 -1.303623386 -0.20084153
#> 10  010  1.806450000 -0.82174254
#> 11  011 -0.688073821 -0.11988800
#> 12  012 -1.889664064 -0.23470142
#> 13  013  0.593381224  1.67636119
#> 14  014 -2.024608511 -1.35688946
#> 15  015 -1.405548714 -0.91882149
#> 16  016  1.617774999  1.17304064
#> 17  017  0.746328883 -0.64944260
#> 18  018 -0.384173084 -1.02983891
#> 19  019 -1.245277655 -1.81651832
#> 20  020 -1.547556736 -0.83331078
#> 21  021 -0.073717752 -0.07384746
#> 22  022 -0.226636304 -0.07285588
#> 23  023  0.539926962  0.04222816
#> 24  024 -0.086667348 -2.42307708
#> 25  025  1.291518211 -0.68604703
#> 26  026  0.616981589  2.57813505
#> 27  027 -1.312327881 -0.24415551
#> 28  028 -0.566680978  0.17826853
#> 29  029  0.802969500 -0.32687544
#> 30  030 -0.723287249 -1.01612187
#> 31  031  0.551109177  0.82000385
#> 32  032  0.821141716 -0.82703352
#> 33  033  1.663698559  1.00051328
#> 34  034  1.905888658  0.32148615
#> 35  035 -0.173029399 -0.93166577
#> 36  036  0.554471872 -0.86964886
#> 37  037 -0.248738075 -1.14123987
#> 38  038  0.713118690  0.14074749
#> 39  039 -0.089334270 -1.03330056
#> 40  040  0.714743611  1.40016925
#> 41  041 -0.714341897  0.31615569
#> 42  042  0.722731338  0.89454492
#> 43  043 -0.736047738 -0.01892174
#> 44  044 -0.162073326 -0.18497339
#> 45  045  0.252427336 -0.77543800
#> 46  046  1.603619683  1.15462130
#> 47  047 -0.339359894  0.60775770
#> 48  048  0.458010079 -0.91518212
#> 49  049  0.236462948  1.10930772
#> 50  050  1.512001951  1.88519362
#> 51  051  0.743960717 -0.08471236
#> 52  052  0.402026502 -1.08738079
#> 53  053 -0.296847676  0.07681631
#> 54  054 -0.253049959  0.46289537
#> 55  055  0.031251400  1.62882779
#> 56  056  0.156621814  0.02818963
#> 57  057 -2.034974974  0.65410335
#> 58  058 -0.685550046 -0.13584818
#> 59  059  0.544892661  0.62667145
#> 60  060 -0.522277060  0.48956648
#> 61  061 -0.589254575 -0.96928707
#> 62  062  2.194453716  1.45589820
#> 63  063 -0.099632096  0.83037036
#> 64  064  0.318630856  0.22910223
#> 65  065  1.604083661  0.12585392
#> 66  066  0.687160679  1.58425051
#> 67  067 -1.922018383  0.72733006
#> 68  068 -0.308144075  0.07793618
#> 69  069  0.674992887  1.86014796
#> 70  070 -0.626128959  0.85239407
#> 71  071  0.098594341 -0.13899681
#> 72  072 -0.825056448 -1.60016081
#> 73  073  0.461455633  0.04078982
#> 74  074 -1.356901489  0.46202340
#> 75  075 -0.304683917 -0.19944833
#> 76  076 -1.008021556 -0.40035574
#> 77  077 -0.311316293  0.01840802
#> 78  078  0.332878169  1.02431075
#> 79  079 -0.081594754 -1.08601108
#> 80  080  0.761846305 -0.67009668
#> 81  081 -0.538477205 -1.82345766
#> 82  082 -1.952249138 -0.61996208
#> 83  083  0.150255012  0.23424847
#> 84  084  0.395057507 -1.43052739
#> 85  085 -0.446172668 -0.06805467
#> 86  086  1.016532228  0.66010873
#> 87  087  0.673050232  1.39426054
#> 88  088  0.180738090  0.72896347
#> 89  089  1.707737680  1.54520513
#> 90  090  1.578607896  0.43243722
#> 91  091  0.168275261  0.18657342
#> 92  092  1.025983870 -1.20864577
#> 93  093  0.173355197  0.68350128
#> 94  094  0.923414136  1.09884301
#> 95  095  0.008500952 -0.50130215
#> 96  096 -0.239770346  0.08468587
#> 97  097  1.395848579  1.18366463
#> 98  098 -0.842317101 -1.14357098
#> 99  099 -0.969784034 -0.84591673
#> 100 100 -0.079968417 -0.54796180