Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5468896
#> Y_2 0.5468896 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.7760120 -0.61875644
#> 2  1.4726254 -0.59121017
#> 3 -2.1198462 -1.59514258
#> 4 -0.3860355 -2.00518247
#> 5 -0.1274568  0.09328099

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.294069625 -0.798922716
#> 2   002  0.819964227 -1.134735410
#> 3   003  0.561837137  1.369470202
#> 4   004 -0.117785325 -0.439177500
#> 5   005 -1.391427075  0.129744420
#> 6   006  0.892426326 -0.802701656
#> 7   007 -0.170144138 -0.673126691
#> 8   008 -1.695468807 -1.485813053
#> 9   009  0.598961404  0.328042161
#> 10  010  0.803916975  0.754749402
#> 11  011  1.682044071  0.623968543
#> 12  012  0.692150223  0.252008884
#> 13  013 -1.069651766  0.824883790
#> 14  014  0.916530052  0.467440261
#> 15  015 -1.417929573 -2.347319938
#> 16  016  0.793640859 -0.534806495
#> 17  017  1.307559523  0.757417915
#> 18  018  0.787871329 -1.027021391
#> 19  019  0.190979237  0.032017250
#> 20  020 -0.124579421 -1.148361464
#> 21  021 -0.766233581 -0.003779057
#> 22  022 -0.745469185 -1.275499082
#> 23  023  2.243563115  1.767810210
#> 24  024 -1.457789495 -0.107480914
#> 25  025  0.122820727  0.969658995
#> 26  026  1.658070298  1.162724102
#> 27  027 -0.438755946 -0.005191228
#> 28  028  2.342842134  1.281139005
#> 29  029  0.871956087  0.813880871
#> 30  030 -0.278236645 -0.242316039
#> 31  031 -0.457023159  1.148622793
#> 32  032  0.389344546 -1.062983242
#> 33  033  1.307782381  1.500686266
#> 34  034  1.358378814  0.861946328
#> 35  035  0.697574511 -0.268748657
#> 36  036 -0.768948617  0.560929602
#> 37  037  0.824887226  0.965600202
#> 38  038  1.745045720  0.935629938
#> 39  039  1.259971995 -0.245163921
#> 40  040 -0.987940836 -2.187215612
#> 41  041  0.142176298 -0.596514501
#> 42  042 -0.525891014 -0.839223755
#> 43  043  0.297244037  0.169094908
#> 44  044  2.495113974  1.289458312
#> 45  045  0.097604715  0.292787264
#> 46  046  0.504137439 -0.533261261
#> 47  047 -0.835662699  1.258131801
#> 48  048 -0.338389571  0.335504807
#> 49  049 -1.447369030  0.938530892
#> 50  050  1.671501692  2.208841601
#> 51  051 -0.490943739 -1.026506221
#> 52  052 -0.356116567 -0.330551236
#> 53  053 -1.289405835  0.073270103
#> 54  054  0.554795662  1.594933982
#> 55  055 -0.154601283 -0.366461341
#> 56  056  0.206582875 -0.191588378
#> 57  057 -0.733006681 -1.568347556
#> 58  058 -1.249435736 -1.933328300
#> 59  059  1.592593582  1.860995389
#> 60  060 -0.384590429  1.203768680
#> 61  061  0.619613847 -0.625576567
#> 62  062  0.257895160 -2.149958980
#> 63  063  0.377179486  1.980076262
#> 64  064 -1.044985943  0.572895834
#> 65  065  0.556062658  0.863332824
#> 66  066 -1.174067920 -0.652532043
#> 67  067  1.524348507  1.875973365
#> 68  068  0.461046377  1.357771853
#> 69  069  0.132199140  0.494625728
#> 70  070 -0.535686522 -0.434308124
#> 71  071 -0.390295179 -0.500783314
#> 72  072  0.467950240  0.315109761
#> 73  073  0.635422776  0.645055489
#> 74  074  1.436366426  0.222663760
#> 75  075  0.305348226  0.133819020
#> 76  076 -1.051561406 -2.109764006
#> 77  077  0.150855160 -0.911859127
#> 78  078  0.804252807  1.139486316
#> 79  079  0.156136269  0.336084243
#> 80  080  0.462058014  1.983397729
#> 81  081 -0.995897699 -0.547441303
#> 82  082 -0.221706122 -0.684755783
#> 83  083  0.813150334 -0.732550538
#> 84  084 -0.644467964 -0.181584655
#> 85  085  1.326599980  1.236230917
#> 86  086  0.374337628  0.388422837
#> 87  087  0.197113132 -0.262994861
#> 88  088 -0.770539094 -0.627405129
#> 89  089 -1.487039411 -1.894898675
#> 90  090  0.002895628  0.459607591
#> 91  091  0.873979107 -0.194648635
#> 92  092  0.220517931  2.002299434
#> 93  093  1.331715217  0.602958677
#> 94  094 -0.631580824  0.737002117
#> 95  095 -0.271877813  0.396318777
#> 96  096  0.759801091  0.110283296
#> 97  097  0.433079782 -2.581222685
#> 98  098  1.487392200  0.065992476
#> 99  099 -0.230824051 -1.018104998
#> 100 100 -0.589296749 -0.967686817

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  0.7669005370 -0.14788599
#> 2   002 -0.6624659620 -0.06098233
#> 3   003  0.4218937314  1.94959893
#> 4   004 -0.3548016859 -0.33749617
#> 5   005 -1.0543391329 -1.21207931
#> 6   006  0.9864259895  1.01463044
#> 7   007 -0.9353844220  0.10663138
#> 8   008  0.2220558517 -0.27484170
#> 9   009  2.3770084610 -0.09883373
#> 10  010  2.5158771267  1.14823089
#> 11  011 -2.0244045820 -0.29474079
#> 12  012  0.2229467001  1.13308499
#> 13  013 -0.5076164122 -1.73386367
#> 14  014 -0.6922093197 -0.56533004
#> 15  015  0.1117817340  0.67965502
#> 16  016 -0.7206153781 -0.14270253
#> 17  017  1.5962558254  2.61435753
#> 18  018  2.6114621822  1.97375527
#> 19  019  0.5088969466  0.74511735
#> 20  020 -0.3525083350  0.80127502
#> 21  021  0.3791847105  0.73029445
#> 22  022  1.8404234432 -0.12979394
#> 23  023  0.0785766347  0.24399376
#> 24  024  1.3596662433  0.41570906
#> 25  025 -0.6870700958 -0.33351495
#> 26  026  0.5833695213  0.09006835
#> 27  027 -1.6137878931 -0.42580016
#> 28  028 -1.3926067071 -1.08449817
#> 29  029  0.4941846899  1.49730754
#> 30  030 -1.6165321639 -0.38526476
#> 31  031  0.8792874182  2.36424001
#> 32  032 -0.0004313114 -0.32703495
#> 33  033 -0.7558363713 -1.25660352
#> 34  034 -0.7507420354 -0.13702409
#> 35  035 -0.4336680032  0.44498701
#> 36  036  0.7891315922  1.78994226
#> 37  037 -1.1546092489  0.23339667
#> 38  038  1.6363273426  0.21645636
#> 39  039 -0.9517967674  1.01390006
#> 40  040  0.5808500323  1.15735743
#> 41  041 -0.6126789963  0.22979720
#> 42  042 -0.0123582015  0.52741382
#> 43  043 -1.6573274987 -0.52034921
#> 44  044 -1.0514029567 -1.31793071
#> 45  045 -0.4565262282  0.46791085
#> 46  046  1.4165390400  0.29085237
#> 47  047 -0.1275013199  0.73727951
#> 48  048  0.0621872581 -1.05400057
#> 49  049 -0.3522567644 -0.01227529
#> 50  050 -0.3125473844 -0.28391985
#> 51  051 -0.5174336603 -0.85165788
#> 52  052 -0.6989170058 -0.53086367
#> 53  053 -0.6091265329  1.15532051
#> 54  054  1.2141160243  1.19991737
#> 55  055  0.4214153600  0.08083195
#> 56  056 -0.9735790725 -0.98819553
#> 57  057 -0.6435179380  0.30227199
#> 58  058  1.2961843505  0.35780171
#> 59  059  0.6606488488  1.14763624
#> 60  060  1.1993737896 -0.07520920
#> 61  061 -0.8610577786  0.18269496
#> 62  062 -0.6910037846 -0.30558269
#> 63  063  1.8416900751  1.11356593
#> 64  064  0.2000659156 -0.07820486
#> 65  065  1.0108714103  0.07734885
#> 66  066  0.1557292470  0.75497843
#> 67  067  2.1736000519  0.96476796
#> 68  068  1.1252463178  1.33602633
#> 69  069  0.1818864384  0.85162907
#> 70  070 -0.0128735546 -1.37872980
#> 71  071  0.5238544749  1.07765988
#> 72  072 -0.4272656008 -0.84744533
#> 73  073  0.9712697001  0.95130338
#> 74  074  1.0149026431 -0.08858127
#> 75  075 -0.7394984493  0.55730172
#> 76  076  0.8584876584  0.07037539
#> 77  077 -0.3544317771 -0.03738079
#> 78  078 -1.6591569980 -1.02519635
#> 79  079  1.5350068314  0.88345800
#> 80  080  0.1163632745  0.12557281
#> 81  081  0.2873719649  0.08487767
#> 82  082  0.2433245505 -0.07220629
#> 83  083  0.1670396004 -0.19827930
#> 84  084  0.9879520515 -0.22566552
#> 85  085  0.5116737999 -1.08820557
#> 86  086  0.0868566757  0.62496625
#> 87  087  0.8956903326 -1.12465084
#> 88  088  0.8932836003 -0.14281634
#> 89  089  0.6770306659  0.59943111
#> 90  090 -0.6849031019 -0.36370422
#> 91  091 -1.0121217450  0.94034537
#> 92  092  0.6807924486  1.09926064
#> 93  093  1.2042286708  0.74400964
#> 94  094 -2.1451869999 -1.11075010
#> 95  095  0.3767113725 -0.04302120
#> 96  096 -0.4447681759 -0.45826775
#> 97  097  1.6708994610  2.33356178
#> 98  098 -0.0897368519  1.60991206
#> 99  099  0.3004030726  1.47598317
#> 100 100 -0.4408222215 -0.41529151