Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4958176
#> Y_2 0.4958176 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1       Y_2
#> 1  0.9729577 -1.091725
#> 2 -0.0105285  1.406239
#> 3 -0.3613931 -1.896470
#> 4 -0.4893257 -1.424189
#> 5  0.7854679  1.028552

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  0.531041979 -0.42698449
#> 2   002 -0.327793414 -0.96088671
#> 3   003  0.987178433  1.54346601
#> 4   004  0.368241499 -0.57844176
#> 5   005  0.420599002  0.38437132
#> 6   006 -1.320780219  0.97579961
#> 7   007  0.014802828 -0.16419653
#> 8   008  1.221104914  0.63410837
#> 9   009  0.302684915  0.79958691
#> 10  010 -0.917158759 -1.08134013
#> 11  011  0.192545251  0.47000305
#> 12  012  1.471232708  1.22944741
#> 13  013  0.330802275 -0.31568819
#> 14  014 -0.977435412 -1.10308621
#> 15  015 -0.086224860  0.49738515
#> 16  016  0.131850601  0.82196200
#> 17  017  1.748382602  0.32268330
#> 18  018  1.028872451 -0.91019050
#> 19  019  0.749138913 -0.40886405
#> 20  020  0.481854070  1.05592641
#> 21  021 -0.516734026 -0.55956646
#> 22  022  1.173891825  1.73546449
#> 23  023 -1.426359051 -1.26717744
#> 24  024 -0.235163405 -0.83644703
#> 25  025  0.962274264  1.36269269
#> 26  026  1.743956034  0.43379083
#> 27  027  0.453805072 -0.15600671
#> 28  028 -0.625200211  0.95623175
#> 29  029  0.637049083  0.05836078
#> 30  030 -1.352469194 -0.10453495
#> 31  031 -0.925990156 -1.16355197
#> 32  032 -0.564925599  0.47652085
#> 33  033  1.478174988  1.50615767
#> 34  034 -1.048245726 -0.71050904
#> 35  035 -0.376353907 -1.15719765
#> 36  036  1.181557936  0.17223403
#> 37  037 -1.710605375 -0.91515009
#> 38  038 -0.303269943 -0.84042560
#> 39  039 -0.412283966 -0.16822339
#> 40  040 -0.506783844 -1.64999916
#> 41  041  1.231611655 -0.16881932
#> 42  042  0.502282792 -0.64569428
#> 43  043 -1.957781167 -1.69233853
#> 44  044 -1.536698905  1.02890223
#> 45  045 -0.469538908  0.92284423
#> 46  046  0.414722576  0.28730240
#> 47  047 -1.581477985 -1.70280881
#> 48  048 -0.990993313 -1.12526951
#> 49  049 -2.648274166 -2.51777502
#> 50  050  0.982920974  1.19152253
#> 51  051 -0.302232229  0.13610998
#> 52  052  2.211359984  1.43161104
#> 53  053 -1.433221783  0.58388624
#> 54  054 -1.280044137 -1.57652577
#> 55  055  0.368429512 -0.78027564
#> 56  056  0.010062176 -1.70522104
#> 57  057 -0.304149609 -1.19405505
#> 58  058  0.546325939  0.43189184
#> 59  059  0.496564531 -0.73084192
#> 60  060  0.059475026  0.12887288
#> 61  061 -1.716127511  0.12910087
#> 62  062 -1.675740719 -1.56877709
#> 63  063 -0.000465282 -0.67093580
#> 64  064  0.207077037  0.31577246
#> 65  065 -1.284520161  0.93476829
#> 66  066 -0.055674581 -0.69517279
#> 67  067 -1.496866383 -0.99718269
#> 68  068  2.001247189  0.11372717
#> 69  069 -0.081360300  0.37057324
#> 70  070 -1.318378612  0.10399258
#> 71  071  0.396093936  1.50893951
#> 72  072  0.550072328  2.44186026
#> 73  073  0.442828180 -0.50785692
#> 74  074  0.975701407 -0.26516815
#> 75  075 -0.866727048 -0.60235405
#> 76  076  0.525856319  0.73333075
#> 77  077  0.967878153  0.51132435
#> 78  078  2.766962430  2.39481147
#> 79  079  0.364632402  0.69782977
#> 80  080  2.036267512  2.96218954
#> 81  081  0.331704850 -0.21441224
#> 82  082 -0.453261119 -1.27544113
#> 83  083 -0.048286360  0.82734759
#> 84  084  0.878821604  0.37623042
#> 85  085  0.410112573  1.14431011
#> 86  086  1.278952011  0.66080648
#> 87  087  0.800747239  0.93903354
#> 88  088  1.109660354 -0.27875340
#> 89  089 -0.425378875 -1.13347861
#> 90  090 -0.166769635  0.23294730
#> 91  091  1.081600751 -0.22640260
#> 92  092 -0.450913156  0.41229179
#> 93  093 -0.270515830  0.06088450
#> 94  094 -1.957634842 -0.63318133
#> 95  095 -0.115390735  1.24359538
#> 96  096 -1.523548625 -0.03781474
#> 97  097  1.280783380 -0.26370089
#> 98  098 -0.612942474 -1.37104066
#> 99  099 -1.357202534  0.30054438
#> 100 100 -0.462630153 -0.04434010

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  0.1913258854 -0.52426491
#> 2   002 -0.4242455925  0.04403400
#> 3   003 -0.7223047105  0.34304238
#> 4   004  0.9815248918  0.98452318
#> 5   005  0.2183064067  0.36867866
#> 6   006  1.2094061123 -0.36862709
#> 7   007 -0.2115592827  0.02021841
#> 8   008  0.3948409859  1.17343571
#> 9   009 -0.8517847002 -2.33152908
#> 10  010 -1.1716287162  0.92130248
#> 11  011 -0.3693114938 -1.35495932
#> 12  012 -0.7542560352 -1.39311586
#> 13  013  0.8166348523  0.11942763
#> 14  014 -0.8577310563  0.09156834
#> 15  015  0.8284193472  0.51807817
#> 16  016  0.8327493930  1.26508928
#> 17  017  0.5240104076 -0.53227782
#> 18  018 -1.0550388898 -0.95047291
#> 19  019 -0.3491574614 -1.27931881
#> 20  020  0.0812490821  1.17565149
#> 21  021  0.2814172764  1.55489604
#> 22  022 -0.6575525389 -1.28583484
#> 23  023  0.8281426116  0.28351616
#> 24  024 -0.2129924857  2.20875154
#> 25  025 -1.0241848305 -1.32902319
#> 26  026 -0.7213132657  0.51245870
#> 27  027  1.1137731881 -0.02349782
#> 28  028 -0.8361813575  0.01486164
#> 29  029 -1.0658643875  0.56384806
#> 30  030  1.1149528041  0.24876006
#> 31  031  1.0401252726  2.40643747
#> 32  032 -1.7633015214 -1.98558252
#> 33  033 -0.0249877859 -0.05895988
#> 34  034  0.3255110661  1.26795924
#> 35  035  0.6112424436  1.19711202
#> 36  036  0.1188691041  1.18531024
#> 37  037 -0.2856987505 -0.29131916
#> 38  038  0.0471913761  2.66838648
#> 39  039 -0.3978411573 -0.83193631
#> 40  040 -0.2691450967  0.30498139
#> 41  041 -0.9574461098 -2.93298690
#> 42  042 -1.2296002673  0.19949420
#> 43  043 -1.2097814638 -1.53639028
#> 44  044  0.1367937032  0.09733933
#> 45  045  1.4823782287  1.87832658
#> 46  046  0.0556603616 -0.51194663
#> 47  047 -1.2794646069 -1.82972275
#> 48  048  1.2522478807 -1.49637442
#> 49  049  1.2714257628  1.39483266
#> 50  050  0.5716535335  1.91393841
#> 51  051  0.8779514823 -0.61233400
#> 52  052  0.0438406653 -0.38228356
#> 53  053  1.8178050633  2.69309555
#> 54  054  0.5314844893  1.40870143
#> 55  055 -0.4236984294 -0.32152982
#> 56  056 -0.5466157984  0.43446505
#> 57  057 -0.5552713908 -0.16083185
#> 58  058 -1.1154126527 -0.87129521
#> 59  059  0.0558085365 -0.53418607
#> 60  060 -1.0192415095 -0.62908041
#> 61  061  0.4388590388  1.36269519
#> 62  062  1.2680040735  0.94909236
#> 63  063 -2.0583744935 -1.51900332
#> 64  064 -0.1664992976  0.09381198
#> 65  065  0.0517838113  0.41119310
#> 66  066 -0.4324032122 -1.34857000
#> 67  067 -1.3007251197 -0.48612372
#> 68  068  0.6497754747  1.57008964
#> 69  069 -1.0886579497  0.46799632
#> 70  070 -0.1249832779 -0.60767196
#> 71  071  1.5453412673  0.78762444
#> 72  072  2.1413467658  0.20423001
#> 73  073  0.0007620419  0.08864835
#> 74  074 -0.1269107612 -1.78648378
#> 75  075 -0.4509893169 -0.37919714
#> 76  076 -0.0247876920 -0.12459734
#> 77  077 -1.5620090919 -1.00278351
#> 78  078 -0.5911093397  0.75918535
#> 79  079 -0.5316077205 -0.47850364
#> 80  080  1.9258152152  0.87563218
#> 81  081 -0.1836428694 -0.67284410
#> 82  082 -0.5587081879  0.41575484
#> 83  083 -2.2965780869 -1.90978344
#> 84  084  0.3514918723  0.19952903
#> 85  085 -0.1736055747  0.30707624
#> 86  086 -1.2124962640 -0.18624893
#> 87  087 -0.6752891136 -1.45049180
#> 88  088 -2.3526045952 -0.44003444
#> 89  089  0.1606408589 -0.38828016
#> 90  090  0.3584293523  0.41223901
#> 91  091  0.5043898637 -1.28657030
#> 92  092  1.1759923418 -0.33827080
#> 93  093 -1.0531849224 -0.94324090
#> 94  094 -2.1374170205 -0.30869461
#> 95  095 -0.3675978777  0.10246981
#> 96  096 -0.9413166251 -0.27607826
#> 97  097 -0.0847884134  0.27817963
#> 98  098  0.2890822212  1.62392187
#> 99  099 -1.3132115424 -0.51942531
#> 100 100  0.1370913872  0.31757350