Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5229234
#> Y_2 0.5229234 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.9215366  0.6186574
#> 2 -1.0425246 -0.1432268
#> 3  0.4867523 -0.4977855
#> 4  0.4714778  2.4328921
#> 5  0.3025958  0.5365987

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.60025453  0.57912035
#> 2   002  0.49620649  0.77481226
#> 3   003 -1.47248241  0.44177737
#> 4   004 -0.97573961 -1.10335420
#> 5   005 -1.80102136  0.25260183
#> 6   006  0.96944571  0.67877573
#> 7   007 -0.59661952  0.52693205
#> 8   008  1.07335277  1.57935126
#> 9   009  1.06636807  1.14586193
#> 10  010 -0.75109757 -1.62238958
#> 11  011 -0.21743833 -1.21290900
#> 12  012  0.22769427  0.97886482
#> 13  013  0.62229898 -1.01582016
#> 14  014  1.43331713  1.28135906
#> 15  015  0.99005939  1.00942352
#> 16  016 -1.22771431 -0.61480917
#> 17  017 -1.77245627 -2.47602020
#> 18  018  1.22917021  2.00693078
#> 19  019  0.21061109  0.60543593
#> 20  020 -1.13505716  1.31620965
#> 21  021 -1.46142092 -1.27643852
#> 22  022 -2.32206663 -1.01418356
#> 23  023  1.45412150  1.84777405
#> 24  024 -0.74911068 -1.01620117
#> 25  025  0.98493355  0.18228073
#> 26  026  0.66017079  0.30404706
#> 27  027  0.43208771  0.61256166
#> 28  028  2.23130742  2.09444055
#> 29  029  0.49557700 -0.09055542
#> 30  030  0.35483805  0.10652256
#> 31  031  1.23126562  0.33148097
#> 32  032 -0.20094435  1.06941723
#> 33  033 -2.22386598 -0.81263352
#> 34  034  2.12781693  0.72787927
#> 35  035 -0.66206471  0.20741089
#> 36  036 -0.40511550  0.13599077
#> 37  037  0.26338370 -0.10703301
#> 38  038 -0.76207065  1.53701501
#> 39  039  2.19439586  0.02818507
#> 40  040  0.87537501  2.14721849
#> 41  041 -1.18650721 -2.22607019
#> 42  042 -0.20200128  0.74927045
#> 43  043  1.72478988  1.34340596
#> 44  044  0.92311072  0.09084770
#> 45  045  0.63989075  0.55985928
#> 46  046  1.26481009 -0.20444191
#> 47  047 -0.59652665 -1.32931070
#> 48  048 -1.17028496 -1.19569978
#> 49  049  0.87766935  0.52100304
#> 50  050  0.89295117 -0.41426941
#> 51  051  2.47929958  2.03602453
#> 52  052 -0.95058274 -1.26260499
#> 53  053 -1.12345682 -1.20051820
#> 54  054  0.40950576 -0.85728003
#> 55  055 -1.36149428 -0.35530768
#> 56  056  0.87875444  0.90454856
#> 57  057 -1.00100925 -0.65100218
#> 58  058  0.35165033  0.70640012
#> 59  059  0.11074665 -0.54321283
#> 60  060 -0.25534703  0.07463014
#> 61  061  1.14986345  1.02258157
#> 62  062  0.88306335 -0.14950127
#> 63  063  0.50668435  0.27347793
#> 64  064 -0.92645268 -1.41230854
#> 65  065  0.91273268 -1.27667060
#> 66  066 -0.04090943  0.05668795
#> 67  067 -0.08263529  0.14402884
#> 68  068 -0.23582940 -1.93342468
#> 69  069 -0.26004267 -0.84896516
#> 70  070 -0.47961717 -0.99669770
#> 71  071  0.73764404 -1.48346493
#> 72  072 -0.26698106 -0.49650764
#> 73  073 -0.45246633 -0.53481444
#> 74  074 -0.95101349 -0.26732202
#> 75  075 -0.33924769  1.14226911
#> 76  076 -1.20225809 -0.28393905
#> 77  077 -0.46753797  1.25691842
#> 78  078 -0.52189156  0.65479129
#> 79  079  0.32334537 -0.84500043
#> 80  080 -1.15030517 -0.57815525
#> 81  081  0.98440022  0.18744314
#> 82  082  1.82386729  0.57080423
#> 83  083 -0.45548545 -0.22108264
#> 84  084  0.46379734 -0.75583554
#> 85  085 -0.61018827  1.19383236
#> 86  086 -0.50405534 -0.39026681
#> 87  087 -0.57761672  0.68378753
#> 88  088  0.29770077  0.19802689
#> 89  089 -1.00279848 -0.58568547
#> 90  090  0.73956659 -0.34441645
#> 91  091  0.45806303  1.16342360
#> 92  092  0.29640489  1.07513342
#> 93  093 -0.21052681  0.61244971
#> 94  094  0.01536464  0.43992745
#> 95  095  0.95433174 -1.12854681
#> 96  096 -3.04712216 -1.77647085
#> 97  097  0.28230784  1.41509001
#> 98  098 -0.78431252 -0.68317731
#> 99  099  1.21558680  0.46627812
#> 100 100  0.30184344 -0.35242455

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.76348184 -0.811492943
#> 2   002  0.10507501  0.497958805
#> 3   003 -0.04810230 -0.656639652
#> 4   004  0.14824883 -0.224827307
#> 5   005  0.02530467 -0.177912678
#> 6   006  0.15786089 -0.356610868
#> 7   007 -0.15368521 -0.008628787
#> 8   008  0.45677852  0.154491726
#> 9   009  0.36737712  0.466985886
#> 10  010 -0.90824143  0.120377937
#> 11  011 -0.78077407 -0.191212468
#> 12  012 -0.09507581  0.799671217
#> 13  013 -0.88998840 -0.837479557
#> 14  014 -0.17138090 -0.195676168
#> 15  015  1.42223448 -0.953974594
#> 16  016 -0.14674591 -0.525180493
#> 17  017 -0.41926499  0.181287782
#> 18  018  0.04765961  0.006172225
#> 19  019  0.26496848  1.379064947
#> 20  020  2.11181100  0.256207207
#> 21  021  1.23144395  0.391583082
#> 22  022 -1.19681785 -1.040770668
#> 23  023 -0.06125779 -2.026010862
#> 24  024 -0.35277487 -1.113616423
#> 25  025  0.15034267  0.093067255
#> 26  026 -2.13501334 -0.768096634
#> 27  027  0.59630500 -1.510897309
#> 28  028 -0.65246075 -0.443281305
#> 29  029 -0.17427612  0.331010285
#> 30  030 -2.80369490 -0.815772321
#> 31  031 -1.45069771 -0.517028801
#> 32  032 -0.22386173  0.835486146
#> 33  033  0.61733867 -0.342760313
#> 34  034  1.67040326  1.171170839
#> 35  035  0.92884627  1.275322153
#> 36  036  0.36483472  0.290736719
#> 37  037 -1.05602226 -0.782629140
#> 38  038 -0.73437274 -1.243602523
#> 39  039  1.92079315  1.923750787
#> 40  040  2.05473449 -0.015631713
#> 41  041  0.84812366  1.584548342
#> 42  042 -0.63946415 -0.544591579
#> 43  043  2.61532219  1.172191699
#> 44  044 -0.40516769  1.025633687
#> 45  045 -2.23214783 -1.393809179
#> 46  046 -0.45677393  0.551757898
#> 47  047 -1.35096931 -0.818304793
#> 48  048  1.21948351  0.263573202
#> 49  049  1.38170320  1.149511500
#> 50  050  0.79403809 -0.293804424
#> 51  051  1.11060024 -0.283515034
#> 52  052  0.04750446 -0.401160920
#> 53  053 -0.92468226 -1.169888727
#> 54  054 -0.08984418 -0.859287550
#> 55  055 -1.12234999 -1.617625302
#> 56  056  1.49237384  2.341255922
#> 57  057 -1.35360519 -0.199760306
#> 58  058  1.35056676  0.654795879
#> 59  059  1.50732688  1.160793273
#> 60  060 -0.50667280 -1.841979935
#> 61  061  0.10878992  0.207137156
#> 62  062  2.29132110  0.574069058
#> 63  063  1.24839282  1.919951782
#> 64  064  1.30631168  0.500320113
#> 65  065  0.56755857  1.373286049
#> 66  066  0.29695563  0.524263427
#> 67  067  0.32166860  0.795219172
#> 68  068 -0.80608705  0.879876536
#> 69  069  0.39257696  0.273604103
#> 70  070  1.30261723  0.942425974
#> 71  071  0.15655478 -0.410900514
#> 72  072 -0.11767429 -0.663071486
#> 73  073 -0.87333181 -2.797867801
#> 74  074  0.62563673 -0.985295226
#> 75  075  0.64724116  0.526828550
#> 76  076  1.85959230  0.534365524
#> 77  077 -1.18791097 -0.404632132
#> 78  078 -1.03895368 -0.006626700
#> 79  079  0.29425884 -0.297120134
#> 80  080 -1.45734031  0.075655493
#> 81  081  1.20471351  2.298683534
#> 82  082 -0.28614420 -1.157952071
#> 83  083  0.85766417  1.045956043
#> 84  084  0.12917328 -0.427791285
#> 85  085 -0.53521382  0.451754067
#> 86  086  1.14752113  0.750246988
#> 87  087 -1.25831357  0.531733740
#> 88  088 -0.24664003 -0.575334827
#> 89  089  1.77370369 -0.623528502
#> 90  090  0.16008321 -0.717982436
#> 91  091 -0.31989093 -0.961428347
#> 92  092 -0.33088998  1.278263548
#> 93  093  0.45652925  0.274252603
#> 94  094 -1.10815867 -1.726829861
#> 95  095 -0.89934790  0.730581169
#> 96  096  0.24982929 -0.715979577
#> 97  097  0.06691753 -0.661334706
#> 98  098 -1.84963012 -0.625992526
#> 99  099  0.37813723  2.507927514
#> 100 100 -0.18493153 -0.980138684