Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4938246
#> Y_2 0.4938246 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  1.0747934  0.07919086
#> 2  0.4461768  0.43026062
#> 3 -1.8713265 -2.68490924
#> 4 -0.7995804  0.32012824
#> 5  0.3502937  0.88754515

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.62460956  0.91986764
#> 2   002  0.08670934  0.80479448
#> 3   003  1.01047615 -0.69899432
#> 4   004  1.06744430  0.80459252
#> 5   005  0.73632063  1.47935295
#> 6   006  1.78572375  0.72690004
#> 7   007  0.79740608  2.58766288
#> 8   008 -0.01015352  0.25800550
#> 9   009  0.86662657  0.94768093
#> 10  010 -1.19896962 -1.26071679
#> 11  011  0.97473751  0.06290807
#> 12  012  1.40896287  2.12957117
#> 13  013  1.60523238  1.22237929
#> 14  014 -0.15845078 -0.65506384
#> 15  015 -1.21181239 -0.18871361
#> 16  016 -1.36279405 -1.43525090
#> 17  017 -0.75861558 -0.07160725
#> 18  018 -0.97561596  0.58717354
#> 19  019  0.23972878 -0.35071858
#> 20  020  1.08704185 -1.11856606
#> 21  021  1.02739124  1.31061353
#> 22  022 -0.49075420 -0.55682505
#> 23  023 -0.78536853 -1.92332539
#> 24  024  0.61513987 -1.03608609
#> 25  025  0.07948232  0.80512690
#> 26  026  0.66658235  1.26379970
#> 27  027  0.85416327  2.58149042
#> 28  028  0.16400068  0.52965903
#> 29  029  0.33092202  1.86548466
#> 30  030 -0.34896718  0.49565895
#> 31  031  0.13726891  1.65178118
#> 32  032 -0.34984438 -1.02495925
#> 33  033 -1.21979486 -0.92754278
#> 34  034 -0.18042990  1.38681859
#> 35  035 -0.10051459 -0.27523014
#> 36  036 -1.26236826 -1.58827797
#> 37  037  1.01383613  0.16291060
#> 38  038 -0.86262180  0.80180980
#> 39  039 -0.10768816  0.16700746
#> 40  040  1.73346828  0.47120839
#> 41  041  2.68712116  1.31279166
#> 42  042  0.70865948  0.46736404
#> 43  043  0.60132627  1.00386182
#> 44  044 -1.16595758  0.70472565
#> 45  045  0.28230608  0.70537854
#> 46  046 -1.60494685 -0.03231315
#> 47  047  0.34283199  0.18754928
#> 48  048  0.79709603 -0.90688160
#> 49  049  0.22394311  1.13323075
#> 50  050  0.29748771  0.31352951
#> 51  051 -0.03165575  0.02352792
#> 52  052  0.40810423 -0.37982929
#> 53  053 -1.36293512 -1.06221239
#> 54  054 -0.06176632 -0.90971131
#> 55  055 -1.42694721 -0.50527301
#> 56  056  0.16453537 -0.36600264
#> 57  057 -2.32923027 -0.04408846
#> 58  058 -1.18956190 -0.89919847
#> 59  059  1.00266591 -1.02671788
#> 60  060 -0.08486096  0.51587890
#> 61  061  2.92169705  1.54309294
#> 62  062 -0.95591678 -0.94904311
#> 63  063  1.04700011  1.77902140
#> 64  064  0.37903401  0.21292239
#> 65  065 -0.09744305 -1.08604110
#> 66  066  0.29307809  0.17829764
#> 67  067 -1.27490849 -0.36529508
#> 68  068 -0.69068379 -0.71988620
#> 69  069  0.57238743  0.30079107
#> 70  070  1.93606608  0.47315168
#> 71  071 -0.46681358 -0.23320257
#> 72  072 -0.85871117 -0.69043431
#> 73  073 -0.23006479 -0.45072378
#> 74  074  0.55408519  0.96814891
#> 75  075  0.08120226 -1.42520668
#> 76  076  1.61020093 -0.43155637
#> 77  077 -0.39460409 -0.99849553
#> 78  078 -0.54647566  0.10663051
#> 79  079  0.49508229 -0.51368642
#> 80  080  1.86747802 -0.11124163
#> 81  081  0.95613356 -1.13383465
#> 82  082 -1.10289377 -0.59697745
#> 83  083 -0.15930594  0.91147408
#> 84  084  0.49672679 -0.27424301
#> 85  085  1.00745575  0.54232657
#> 86  086  1.43643135  0.46413570
#> 87  087  0.05797021 -0.73670857
#> 88  088  0.39054525 -0.40267845
#> 89  089  0.82738238 -0.19773170
#> 90  090  1.00213606  0.09237679
#> 91  091 -0.39159581 -0.33955388
#> 92  092  1.58459504  2.34845043
#> 93  093  0.03666710  0.02360600
#> 94  094 -1.65207112 -1.14978219
#> 95  095  0.40754194 -0.12987206
#> 96  096 -0.69584337  0.07340432
#> 97  097  0.38939948  0.64606447
#> 98  098 -0.45559969  0.47720427
#> 99  099 -1.20868356 -0.13575121
#> 100 100  1.34252156  0.38500425

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  1.064604983  0.171039450
#> 2   002 -0.144316941 -0.834931495
#> 3   003 -0.213083675 -0.189403160
#> 4   004  0.593141238  1.945066491
#> 5   005 -0.332758081  0.890678686
#> 6   006  1.039279216  1.269288045
#> 7   007  1.250626858  1.486246609
#> 8   008 -1.307720725 -0.945407814
#> 9   009  0.854839650 -0.241988805
#> 10  010  0.036743340  0.390338604
#> 11  011  0.558551768  0.426929206
#> 12  012  0.020791695 -0.928057369
#> 13  013  0.187203361  1.538716074
#> 14  014 -1.106547207 -0.396733523
#> 15  015 -1.246515721  0.647961764
#> 16  016 -0.955719376  0.519513315
#> 17  017 -0.750522978  0.006313162
#> 18  018  0.017020502  0.421533815
#> 19  019 -0.463142918 -2.369609547
#> 20  020  0.917873114  1.392334570
#> 21  021 -0.772565647 -1.362148771
#> 22  022 -0.753417141 -0.616149150
#> 23  023  0.100409194 -0.035431731
#> 24  024  0.612146077 -0.794648987
#> 25  025 -1.136240642 -0.499401642
#> 26  026 -1.033547858 -1.611480365
#> 27  027 -0.103924552 -0.876079228
#> 28  028  0.801421716  0.662819809
#> 29  029  0.442941900 -0.069021572
#> 30  030  0.001143108 -0.005298559
#> 31  031 -0.647479079 -0.633257453
#> 32  032 -0.116684000  0.871825839
#> 33  033 -2.047675671 -0.464228866
#> 34  034  0.732345097 -0.696364146
#> 35  035 -0.617772484 -0.637346847
#> 36  036  0.344445899  0.913477993
#> 37  037 -1.191081827 -0.947342328
#> 38  038 -1.260927653 -0.925404415
#> 39  039  2.254243480  1.700089744
#> 40  040 -0.514528758  0.848215898
#> 41  041 -0.403171693 -1.232343770
#> 42  042 -0.070884791  0.282410972
#> 43  043 -0.493148576  0.709286914
#> 44  044 -0.534970978 -0.434822194
#> 45  045  1.695357643 -0.705106216
#> 46  046 -0.167482916  1.026694287
#> 47  047  0.496232396  1.261787381
#> 48  048  1.141145252  0.524608462
#> 49  049 -1.475976264 -1.502283019
#> 50  050  0.590881678  0.662507752
#> 51  051  0.615128264  0.593439607
#> 52  052 -1.035088709 -0.874574866
#> 53  053  0.618232434 -0.716627064
#> 54  054 -0.070147782 -1.163034538
#> 55  055 -0.113073874 -0.279322121
#> 56  056 -0.281541942  0.359930586
#> 57  057  0.746012501  0.984545467
#> 58  058  0.640154296  0.602035075
#> 59  059 -0.375778464  0.037500343
#> 60  060  0.434998239 -0.079863802
#> 61  061  0.557452171  0.234383868
#> 62  062  0.423999917 -0.178497453
#> 63  063 -0.189135863  1.336994789
#> 64  064 -1.045818962 -0.465008810
#> 65  065 -0.495841038  0.421870513
#> 66  066  2.122774218  1.263858966
#> 67  067  0.831869158  0.708367783
#> 68  068  0.427361444  0.777787290
#> 69  069 -1.679944156  0.500319932
#> 70  070  0.753553956 -0.465147526
#> 71  071 -1.909399624 -1.160486952
#> 72  072 -0.926250183  1.719862974
#> 73  073 -0.439146842  0.446242498
#> 74  074  0.693366122  1.100714598
#> 75  075 -0.772724012 -0.209334569
#> 76  076 -0.602117930  1.194104862
#> 77  077  1.234797857  0.327276173
#> 78  078 -1.076211182  0.405007681
#> 79  079  1.031013979  0.083005256
#> 80  080  1.307019269 -0.922019413
#> 81  081 -0.096000842 -0.940293609
#> 82  082  0.490948561  0.195491807
#> 83  083  0.010221039 -0.270426528
#> 84  084  0.697340167  1.770753200
#> 85  085  1.885115938  0.266948273
#> 86  086  0.868587293  1.138464184
#> 87  087  0.689511487  0.407275353
#> 88  088 -1.837071982  1.023415579
#> 89  089  1.294739376  0.248580034
#> 90  090 -1.230001183 -2.183651302
#> 91  091  1.867212733 -0.067989731
#> 92  092 -1.228678209  0.070706403
#> 93  093 -0.511151722  0.834158917
#> 94  094  1.711218565  0.441967048
#> 95  095 -2.089377755 -1.003620399
#> 96  096 -0.816669543 -1.805476683
#> 97  097 -3.306437660 -1.466725846
#> 98  098 -1.318905969 -1.655819828
#> 99  099  0.140690919  1.227187334
#> 100 100 -0.339531060  1.510648736