Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5185084
#> Y_2 0.5185084 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>            Y_1        Y_2
#> 1  0.711932864  0.8242055
#> 2  1.207086411  1.1389246
#> 3 -1.201065032 -0.8801893
#> 4 -0.163453970 -1.5406451
#> 5  0.006791071 -0.2301846

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.542742946 -1.11266780
#> 2   002  1.826018283  1.55487741
#> 3   003 -0.328793013 -0.60189425
#> 4   004 -1.406429089  0.15509118
#> 5   005 -0.452573428  0.98668450
#> 6   006 -0.758647290  0.76684812
#> 7   007 -0.069971117  0.82607421
#> 8   008  0.016903555 -0.20962947
#> 9   009 -1.153596579 -0.57519564
#> 10  010  0.610547803 -0.04156270
#> 11  011  0.577922683 -1.00417568
#> 12  012  0.055566875 -0.57969899
#> 13  013 -0.183885654 -0.81338243
#> 14  014  1.124027470  0.07827500
#> 15  015  2.611802147  0.46375131
#> 16  016 -0.328615467  0.58950629
#> 17  017 -1.499071930  0.71673366
#> 18  018 -0.742289245 -0.55820007
#> 19  019  1.685962264  1.73945022
#> 20  020  1.398107527  1.50073988
#> 21  021  0.086232920  1.49233722
#> 22  022  0.180431784  0.25673817
#> 23  023  0.653094446  0.02638343
#> 24  024  0.063093307 -0.55497565
#> 25  025 -0.682730246  1.13651244
#> 26  026 -0.007855021  0.39422643
#> 27  027 -1.234442135 -1.31079747
#> 28  028 -0.409113065 -0.04581839
#> 29  029  0.994719939  0.66603072
#> 30  030 -1.364353539 -0.56102706
#> 31  031 -0.183592041 -0.37956807
#> 32  032  0.558533160 -0.51225216
#> 33  033  0.509663487 -1.35812915
#> 34  034  0.446641608  0.47495421
#> 35  035 -0.717806954 -0.95684955
#> 36  036  1.333108993  0.34795521
#> 37  037  0.031719753  1.39843426
#> 38  038 -0.308356894  1.87803009
#> 39  039  0.325854277  0.46058524
#> 40  040 -0.804878931 -0.77360456
#> 41  041  1.940017597  1.03055665
#> 42  042 -0.454812047  0.44080849
#> 43  043 -0.937852782 -0.32007853
#> 44  044 -1.665958115 -1.41089152
#> 45  045 -0.401222591 -1.75742024
#> 46  046 -0.284743105 -0.54022643
#> 47  047  0.421926552  1.42521667
#> 48  048  1.132286129  2.27880429
#> 49  049 -0.176941055 -1.32188240
#> 50  050  0.716998162 -1.52648469
#> 51  051  0.163610797  0.50309014
#> 52  052  0.983216646  0.85601815
#> 53  053 -1.277011535  0.43654693
#> 54  054 -0.206046099  0.41795560
#> 55  055 -0.194052844 -0.85934196
#> 56  056  0.241597321  0.29218900
#> 57  057  0.671579958  0.06452798
#> 58  058 -0.050318378  1.43225496
#> 59  059  2.378592958  1.48069419
#> 60  060 -0.995596852  1.04404410
#> 61  061 -1.142748088  0.13935262
#> 62  062  0.859038999  0.38945243
#> 63  063 -0.999838546 -1.13568400
#> 64  064  0.367291918  0.88080144
#> 65  065 -1.014739410 -0.17591067
#> 66  066 -1.361367726 -0.75762059
#> 67  067 -0.058988033 -1.78433699
#> 68  068 -2.134916421 -1.58912467
#> 69  069  1.233051287  0.55217387
#> 70  070 -0.086631202 -0.22305749
#> 71  071  0.661355416 -0.10855754
#> 72  072 -0.129717500  0.68928286
#> 73  073 -2.392286958 -1.08695914
#> 74  074  0.056426990  1.33821713
#> 75  075  1.496307506  0.48466671
#> 76  076  2.356983685 -0.12433306
#> 77  077 -0.020112885  0.86382880
#> 78  078  0.499075183 -0.33497259
#> 79  079 -0.330181776  0.18341947
#> 80  080 -0.035647650 -0.08140013
#> 81  081  1.518980030  0.19068559
#> 82  082  0.731811004 -0.31024911
#> 83  083  1.007770115  0.63956184
#> 84  084 -0.467854394  0.48841390
#> 85  085 -0.828675668 -1.70948697
#> 86  086  2.097562773  1.35092780
#> 87  087  1.906349452  0.22448848
#> 88  088 -0.186279264 -0.88655865
#> 89  089 -1.214524118  0.16325103
#> 90  090 -0.131760897 -0.95633816
#> 91  091 -1.449274824 -0.51316814
#> 92  092 -1.318017084 -1.21358745
#> 93  093 -0.791765299 -1.35904201
#> 94  094 -0.363213236 -0.07079250
#> 95  095  0.894806009 -0.34897830
#> 96  096  1.118799363  0.23196334
#> 97  097 -1.454708738 -0.85499451
#> 98  098  0.619484040  2.76470522
#> 99  099  0.273832053 -0.14068675
#> 100 100 -1.311153752 -0.78801053

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.09557433 -1.7955635363
#> 2   002 -0.56351135  0.6919575822
#> 3   003 -0.13007813 -0.2234419894
#> 4   004 -1.92080837 -1.1789832151
#> 5   005 -1.30740761 -1.1836292582
#> 6   006  0.44304461  1.1349521505
#> 7   007 -0.82622115 -0.6252491673
#> 8   008  0.64818026 -0.1481934959
#> 9   009  0.38596155  0.3211154693
#> 10  010  1.16542820 -0.6385203566
#> 11  011  1.64387485  1.6630809827
#> 12  012  1.83776197  1.2170221757
#> 13  013 -0.28364821  0.0777381147
#> 14  014 -0.35524655 -1.2949444931
#> 15  015  1.54191037  1.0124122040
#> 16  016 -1.07354348 -0.9600215705
#> 17  017 -1.09262193 -1.0966800940
#> 18  018 -0.25425354 -1.1845587821
#> 19  019 -1.05517553  0.7796085817
#> 20  020  0.20618196  0.9557322754
#> 21  021 -1.48357729  0.9779951253
#> 22  022 -1.04155244  0.6732181506
#> 23  023  1.41372769  0.6871887122
#> 24  024 -0.93633552 -0.0495055445
#> 25  025 -1.03204262  0.5616113009
#> 26  026 -2.47851596 -1.9706694598
#> 27  027 -0.41229536  0.5823300720
#> 28  028  0.66595610 -1.0812856284
#> 29  029 -0.92286604 -0.4182592602
#> 30  030  0.94666422  0.4764015886
#> 31  031 -0.26165481 -0.7299138181
#> 32  032  1.38487167  0.6091058652
#> 33  033 -1.08364276 -1.2322848278
#> 34  034  1.34585699 -0.6143773200
#> 35  035  1.11715297  1.8884417385
#> 36  036 -1.10369138  1.2122799662
#> 37  037 -0.52547681 -1.1589337784
#> 38  038 -2.17641749 -0.9032249522
#> 39  039  1.38277251  0.6932474806
#> 40  040  1.31741529  0.0966206584
#> 41  041 -0.72833806 -0.6901212662
#> 42  042  0.19953574  0.9432506932
#> 43  043  0.35653509  0.6362850296
#> 44  044 -1.49503907 -0.4208091811
#> 45  045 -0.35898593 -0.7037068577
#> 46  046  0.24396959  0.4234157692
#> 47  047 -0.71587479 -1.1153919999
#> 48  048 -0.98786938  0.0533449611
#> 49  049  0.23589530 -0.8791448745
#> 50  050 -0.96200149  1.0684737573
#> 51  051 -1.24938712 -1.3885378681
#> 52  052  1.28591115  0.0007445026
#> 53  053 -0.39785856  1.0423054388
#> 54  054 -0.45055140 -0.6389186644
#> 55  055  0.72279374  0.4982979228
#> 56  056  0.04604346 -0.1103067143
#> 57  057  1.47165960  0.8816541364
#> 58  058  0.68915868  0.9150523834
#> 59  059 -1.19546005 -0.9706242477
#> 60  060 -1.25978210 -0.8513865519
#> 61  061 -0.52800039 -0.5422908708
#> 62  062 -0.78174237 -1.5739050609
#> 63  063 -0.59952831  0.0926019900
#> 64  064 -0.33389989  0.1716317713
#> 65  065  0.47923367 -0.1243293108
#> 66  066  0.78552570  0.4667462757
#> 67  067  1.57254834  2.4038254024
#> 68  068 -0.13735784 -0.3948843718
#> 69  069 -1.60248642 -0.8745825626
#> 70  070 -0.70566760  0.5230667197
#> 71  071  0.60456508  1.8077036399
#> 72  072 -1.04837624  0.3645003460
#> 73  073  1.25868328  0.8099972293
#> 74  074 -1.63811235 -1.8617456907
#> 75  075 -1.66217933 -0.8943164553
#> 76  076  0.55960782  2.0550101827
#> 77  077 -0.75122622  1.7677611395
#> 78  078 -1.23808110 -0.7757717523
#> 79  079  0.37892813  1.2470002087
#> 80  080 -0.83069686  1.0094189789
#> 81  081 -0.68963802 -1.5603253655
#> 82  082 -2.21225379 -2.1337581082
#> 83  083 -1.69706904 -1.0169373679
#> 84  084  2.25530082  1.2824731911
#> 85  085  0.90627400  0.6523364814
#> 86  086  1.79398172  0.6431693559
#> 87  087 -0.53019322 -1.0304064872
#> 88  088 -1.70607179 -0.2415268551
#> 89  089 -0.58081153  2.5033887536
#> 90  090  1.00395170  0.9931795481
#> 91  091  1.94188000  0.7584731112
#> 92  092  0.22369597  0.2083933792
#> 93  093 -1.00292633 -0.1606495617
#> 94  094 -0.39924593 -0.0334863393
#> 95  095 -0.87130020 -1.7443559704
#> 96  096 -0.81209377 -0.9028681014
#> 97  097 -1.27177978 -0.9985911225
#> 98  098 -1.71386325 -0.1777855462
#> 99  099  1.30960314  0.1736250647
#> 100 100 -1.57092299 -1.4316308953