Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5058804
#> Y_2 0.5058804 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.3254735  1.5370434
#> 2 -0.5353834 -0.1171319
#> 3 -0.2022422 -0.8963170
#> 4  0.6992215  1.0904163
#> 5  0.7530224 -0.3493890

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1           Y_2
#> 1   001  0.524987355 -0.3775661791
#> 2   002  0.069379332 -0.8452690833
#> 3   003  0.649775210  0.0723053713
#> 4   004  0.158111808  1.0844269467
#> 5   005 -1.373467783 -0.0005414503
#> 6   006  0.779113417  0.6189550747
#> 7   007 -1.822608658 -0.7051231052
#> 8   008 -1.115987157 -1.6574696815
#> 9   009 -1.078458905 -1.6213687303
#> 10  010  1.099331944  0.0705551908
#> 11  011 -1.037510090 -0.3666554744
#> 12  012  1.140731190  0.8807588671
#> 13  013  0.004487538 -1.8542576353
#> 14  014 -0.595795565 -0.3945194741
#> 15  015 -0.397526459  1.1347919943
#> 16  016  1.409294796  0.4931433049
#> 17  017 -1.632830983 -0.1042281409
#> 18  018 -0.858333834 -0.4315745851
#> 19  019 -0.722748496 -2.2218429394
#> 20  020 -0.755633587  0.7394282006
#> 21  021  1.528215713  1.1302512781
#> 22  022 -0.777880495 -1.4417834471
#> 23  023 -1.606474442 -0.0662626200
#> 24  024  0.642800113  1.2878815765
#> 25  025 -1.131534021 -1.2157415552
#> 26  026  1.715321400  0.1520599282
#> 27  027 -0.946915674  0.3823661833
#> 28  028 -0.670769784 -0.4436083320
#> 29  029  0.803485164  0.0728107367
#> 30  030  0.451680069  1.1936431944
#> 31  031 -1.405009798 -0.8708687683
#> 32  032 -0.313344670 -0.8747271006
#> 33  033  0.336915957  0.7012117297
#> 34  034 -0.717945570 -0.9500838015
#> 35  035  0.975285241  0.1691665397
#> 36  036  0.151133881  0.7224132937
#> 37  037  0.534254943 -0.2119783006
#> 38  038  0.091107924  0.2374541046
#> 39  039 -1.270531721  0.4096052746
#> 40  040 -0.478474573 -1.4676072017
#> 41  041 -1.156174380 -1.0305806688
#> 42  042  0.274904852 -0.8761672818
#> 43  043 -0.909799026  0.6292112329
#> 44  044 -0.823542846 -0.2855767713
#> 45  045 -0.022803649  0.0374208629
#> 46  046 -0.199862079  0.3247517500
#> 47  047  0.292411015  0.0834019869
#> 48  048 -0.702322329 -0.4262086284
#> 49  049  0.030058054 -0.2941486506
#> 50  050  1.042128354  0.9356554985
#> 51  051  1.102818957 -0.4251691102
#> 52  052 -1.241355010 -0.1133054249
#> 53  053  0.239776171  0.4731427223
#> 54  054 -0.800075124 -1.6859736789
#> 55  055 -0.107768096 -0.0071606081
#> 56  056  1.626058965  0.4589987863
#> 57  057  1.084904997 -0.2491922067
#> 58  058 -0.259215298 -0.1677822539
#> 59  059  0.909900011  1.2209476520
#> 60  060 -0.967787167 -1.7947582867
#> 61  061  0.232918710  1.9151255729
#> 62  062 -0.057045289  0.9903980244
#> 63  063  1.582372189  1.3866384761
#> 64  064  1.501539476 -0.1444158373
#> 65  065  1.299287132  1.5934452764
#> 66  066 -0.483943665 -0.7401000086
#> 67  067 -0.047798433 -0.2584579734
#> 68  068 -0.519283199  0.5259323753
#> 69  069 -0.095680526 -0.2430226548
#> 70  070 -1.128451895 -1.3329126674
#> 71  071  0.092179405 -0.9002169917
#> 72  072 -1.667736149 -2.4690097574
#> 73  073 -0.456987089  0.2647555516
#> 74  074  0.637608778  0.6060984197
#> 75  075  0.463306106 -0.2396663865
#> 76  076  1.045500174  0.3493680009
#> 77  077 -2.228375902 -1.5657102306
#> 78  078 -1.914670617 -1.0999908469
#> 79  079 -1.184698938 -0.3825748886
#> 80  080 -1.088785656 -0.4055734227
#> 81  081  1.591570960  1.0200356873
#> 82  082  0.823237578  0.7063034277
#> 83  083 -0.884205778 -1.1611189534
#> 84  084 -0.032646645  0.1026110586
#> 85  085 -0.069476060 -0.1759337650
#> 86  086  0.050828474  1.1609975986
#> 87  087  0.924522308 -0.7427141312
#> 88  088  0.154692479  0.1439466713
#> 89  089  0.325272732 -1.3395192168
#> 90  090  0.974331071  0.1866480281
#> 91  091  2.286798683  2.0010994641
#> 92  092 -0.379463370 -0.3596524990
#> 93  093 -0.135176469  0.8721622696
#> 94  094 -1.496730724 -0.9733350862
#> 95  095  1.083399619  0.2365323024
#> 96  096 -0.403035106 -1.2956494971
#> 97  097 -1.352089935 -0.9731165162
#> 98  098  1.595572567  0.9462724623
#> 99  099 -0.310919057 -0.6529682961
#> 100 100  0.496274390  0.9012100012

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.10774398  1.201921003
#> 2   002  0.06967023 -0.235744948
#> 3   003  1.45936488 -0.703141030
#> 4   004  0.35633922 -0.479459591
#> 5   005  0.90087826  1.047988466
#> 6   006 -0.69033312 -0.288710711
#> 7   007 -0.77848910 -0.784023294
#> 8   008  0.57889368 -0.379035087
#> 9   009 -0.13312782  1.595568242
#> 10  010 -1.77874582 -0.420068522
#> 11  011 -0.71898912 -0.219607466
#> 12  012  0.59120074  1.557852740
#> 13  013 -0.47352463 -0.061460308
#> 14  014 -0.67246430 -1.378818154
#> 15  015 -0.20060477 -0.295135462
#> 16  016 -0.47547272 -1.099727248
#> 17  017  0.45455459  0.353912737
#> 18  018  0.69060970  0.684570591
#> 19  019 -0.98716876 -1.485290558
#> 20  020 -0.85878833  0.079039365
#> 21  021 -0.05639408  0.645892051
#> 22  022 -3.38888153 -0.309999562
#> 23  023  1.26054446  1.462598662
#> 24  024 -0.78242142 -0.592102333
#> 25  025  0.66991071  0.127558030
#> 26  026  1.39668006  0.903694053
#> 27  027 -1.27334218 -0.884233902
#> 28  028 -1.05973097 -0.888901893
#> 29  029 -1.10877905 -0.524091670
#> 30  030 -0.99650306  0.028378617
#> 31  031 -0.44722555  0.608190009
#> 32  032  0.25716075  0.144622062
#> 33  033 -1.60847410 -3.125117339
#> 34  034  0.45800856  2.269792618
#> 35  035  0.92214334 -0.359875597
#> 36  036 -0.12795978 -0.124352236
#> 37  037  0.61613021  1.094237456
#> 38  038 -0.20339508  0.994802723
#> 39  039 -0.22120163  0.311488149
#> 40  040 -0.92910655 -0.073353911
#> 41  041 -0.33269264  0.411342263
#> 42  042 -0.99763138  0.048972126
#> 43  043 -0.71083694  0.517145737
#> 44  044  1.73296428 -0.572934307
#> 45  045  0.01924895  0.956715561
#> 46  046  0.75614725  1.562601725
#> 47  047  1.27640546 -1.073779145
#> 48  048 -0.79948666  0.874784982
#> 49  049 -0.45267191  0.237335391
#> 50  050  0.21267886  0.852191422
#> 51  051  0.46774422 -0.067220493
#> 52  052 -1.32541433 -1.317654628
#> 53  053 -0.22986577 -1.653482073
#> 54  054 -0.22676621 -0.414795787
#> 55  055  0.37099575  0.485519054
#> 56  056  1.72513297  0.743787152
#> 57  057 -1.11147601 -0.973031122
#> 58  058  0.28654751 -0.249733885
#> 59  059 -0.14935432  1.192851750
#> 60  060 -1.25843169 -0.841926522
#> 61  061 -0.68695299 -1.140815867
#> 62  062  0.58517109 -0.014987303
#> 63  063  0.45272775  1.082294806
#> 64  064 -0.12484169 -1.326847043
#> 65  065  2.03439098  1.930710341
#> 66  066  0.83906079  1.391919957
#> 67  067  1.38460220  0.460884770
#> 68  068 -0.60905218 -0.364318877
#> 69  069 -0.19366253  1.563499387
#> 70  070  0.77141124  0.888436510
#> 71  071 -1.36462215 -0.904404046
#> 72  072  0.33083335 -0.342547679
#> 73  073 -1.73451645 -0.909789966
#> 74  074  0.25738278  0.004019553
#> 75  075  0.08770915 -2.386102753
#> 76  076  0.04734720  0.277839379
#> 77  077 -1.74732896  1.109365069
#> 78  078  0.84589634 -0.265573925
#> 79  079  0.74208982 -0.132550486
#> 80  080 -1.27748277 -1.821176103
#> 81  081 -0.62413548 -0.471006036
#> 82  082 -0.69195584 -0.269851156
#> 83  083 -0.35830130 -1.608972825
#> 84  084 -0.21178260 -0.887590574
#> 85  085  0.16822596 -1.570724372
#> 86  086 -0.57267239  1.352272087
#> 87  087 -0.34816457  1.055146067
#> 88  088 -0.66579886  0.823288334
#> 89  089 -2.16719869 -0.273525143
#> 90  090 -0.89664888 -0.123565407
#> 91  091 -0.21758709 -0.396401926
#> 92  092  0.74578809  0.077716812
#> 93  093  0.34330256  0.714661135
#> 94  094 -1.28279604  0.587959475
#> 95  095 -1.25081132 -0.549507777
#> 96  096 -0.45827390 -0.649355349
#> 97  097  2.25738813  1.305055789
#> 98  098  0.82969910  1.802570573
#> 99  099 -0.93361072 -0.689683098
#> 100 100 -1.46825353 -1.522989232