Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.496651
#> Y_2 0.496651 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.5233165 -0.25752990
#> 2 -0.5693504 -0.01854146
#> 3 -1.9254433 -0.22754211
#> 4 -0.1133860 -0.18168887
#> 5  1.5660500  2.16055044

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.03567729  0.39785841
#> 2   002 -0.27157521 -0.93516913
#> 3   003 -0.08587892  0.11581177
#> 4   004  0.81524292  0.61056166
#> 5   005  0.77410623 -0.73817348
#> 6   006 -0.41242125  0.46657672
#> 7   007 -0.25222736 -0.29721908
#> 8   008  0.91251255 -0.09804849
#> 9   009 -0.43297447  0.58993053
#> 10  010 -1.12033800 -1.03273584
#> 11  011  0.83347622  0.93283581
#> 12  012  0.92221904 -0.91018011
#> 13  013 -0.08137285  0.06806345
#> 14  014  1.38967753  1.10520115
#> 15  015  0.46373617 -0.63184493
#> 16  016  0.13418967 -1.98809009
#> 17  017 -0.01852526  1.21766768
#> 18  018 -0.28657917 -1.16036863
#> 19  019 -0.17150325 -0.52072833
#> 20  020 -0.34616621  0.03294566
#> 21  021 -2.40710408  0.54600961
#> 22  022  0.18607302  0.06150989
#> 23  023 -2.03621513 -0.16384352
#> 24  024  0.85692000  0.91387360
#> 25  025  0.63279421  0.68495418
#> 26  026 -0.41811414  0.56725681
#> 27  027  0.20107893  1.49601657
#> 28  028 -0.35956924 -1.15925710
#> 29  029  0.77931826  0.12729226
#> 30  030 -0.10778546  0.54272894
#> 31  031 -1.15512882 -0.68036634
#> 32  032  1.80074383 -0.97395170
#> 33  033 -0.64359768  1.39434702
#> 34  034 -1.20421130 -0.11673002
#> 35  035  0.72788875  1.19596570
#> 36  036  0.51996406  0.26050857
#> 37  037  0.33830958  0.32113662
#> 38  038 -0.33624052  0.76806722
#> 39  039  0.39528128  0.71250101
#> 40  040 -1.66606985 -2.39447324
#> 41  041  0.95485993  0.64577152
#> 42  042 -0.43935810 -1.30232940
#> 43  043 -0.50265940 -0.69737324
#> 44  044 -0.39875872 -1.01543749
#> 45  045 -2.10683164 -1.59292957
#> 46  046  0.61882877  0.43534154
#> 47  047  1.59736819 -0.31362570
#> 48  048 -1.11744079 -1.90855624
#> 49  049  0.20414120 -0.56661737
#> 50  050 -1.87115258 -1.11845103
#> 51  051 -2.25837290 -0.43011482
#> 52  052  0.57738279  0.82954949
#> 53  053 -1.95679886 -2.42911339
#> 54  054  1.06895209  1.17883864
#> 55  055  0.05820766 -0.22035894
#> 56  056  1.03789873 -0.66676082
#> 57  057 -1.09295721 -2.32690736
#> 58  058  0.34241967 -1.08348671
#> 59  059 -1.66814262  0.44030282
#> 60  060  1.67052636  3.39133764
#> 61  061  1.37503177 -0.26161250
#> 62  062  1.08856535  1.05980866
#> 63  063 -0.23678638  0.29581615
#> 64  064  0.13212748 -1.00326410
#> 65  065 -0.05834351  0.64052459
#> 66  066 -0.37557925  1.42567031
#> 67  067  0.15593348  1.68806563
#> 68  068 -0.73285246  0.14806648
#> 69  069  0.58535110  0.20712404
#> 70  070  0.25886290  0.92501830
#> 71  071 -0.56311424  0.79880944
#> 72  072 -1.65882709 -0.45443910
#> 73  073 -1.76596102 -0.54609276
#> 74  074  0.09469606 -0.80879005
#> 75  075  0.84101665  1.47223711
#> 76  076 -0.70614843 -0.56229770
#> 77  077 -0.32047432 -0.90661766
#> 78  078 -0.18253204  1.14220106
#> 79  079 -0.72629688 -0.80850682
#> 80  080  0.10610567  0.41148993
#> 81  081 -0.30854005 -0.10616982
#> 82  082  0.94027629  0.15490710
#> 83  083  0.12795437  0.05574265
#> 84  084  1.36454655  0.42877593
#> 85  085  0.35975601  1.22574763
#> 86  086 -1.17843109  0.92846038
#> 87  087  1.68426670  0.60813816
#> 88  088  1.28757073  3.40240679
#> 89  089 -0.36927809  0.48078740
#> 90  090 -0.59297258  1.23460560
#> 91  091 -0.47573082  0.51019815
#> 92  092 -1.94858949 -0.72996115
#> 93  093 -0.50457026 -0.18980447
#> 94  094 -2.05693983 -3.10197765
#> 95  095  0.03210108 -0.11427201
#> 96  096  0.05213122  1.23094804
#> 97  097 -0.43272625  1.00371856
#> 98  098  1.85516825  0.39389017
#> 99  099 -0.29661078  0.34888860
#> 100 100  1.38010880  1.00833287

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.483751138  0.46612147
#> 2   002 -1.370364708 -1.23047472
#> 3   003 -1.127611867 -0.04728686
#> 4   004 -1.128587298 -0.93912405
#> 5   005 -1.020489349 -1.15594205
#> 6   006 -0.082821572 -1.07571628
#> 7   007 -0.510987715  1.83665467
#> 8   008  0.320563675 -0.10466295
#> 9   009 -0.652023967 -1.24249183
#> 10  010  2.245672507  1.48552500
#> 11  011  1.047438766 -1.20040365
#> 12  012 -1.519870202 -0.31782961
#> 13  013 -0.680285223 -0.94655411
#> 14  014  0.854394257  0.38432373
#> 15  015 -2.544656006  0.38416106
#> 16  016 -3.258772967 -2.62826095
#> 17  017  0.431886540  0.40396838
#> 18  018 -0.860088296 -0.70912403
#> 19  019  0.938740451  0.55180277
#> 20  020  0.322899165  0.29561172
#> 21  021  0.477143906  1.56142046
#> 22  022  0.814877594  0.52628159
#> 23  023 -0.486079839  0.15458353
#> 24  024 -2.270789579 -1.36617938
#> 25  025  0.703519970  1.71149378
#> 26  026  1.332130585 -0.32427830
#> 27  027  1.436423984  1.24516856
#> 28  028 -0.625978181 -0.82856466
#> 29  029  0.196032915 -0.27846433
#> 30  030  0.278370451 -0.90304128
#> 31  031  0.564420766  1.13332456
#> 32  032  0.265496945  0.62267759
#> 33  033 -0.104095900 -0.79734382
#> 34  034 -2.428453023 -0.45212169
#> 35  035  1.688587620  0.42803050
#> 36  036 -0.596953730 -2.72314507
#> 37  037 -0.831698485 -0.17847100
#> 38  038 -0.421897678  1.92750844
#> 39  039  0.235775572  0.01673852
#> 40  040 -0.701065855  1.58498174
#> 41  041  0.322144267 -0.58164412
#> 42  042  0.780603660  0.20690938
#> 43  043  0.692982521 -0.88087917
#> 44  044  0.040757547 -0.28421438
#> 45  045  1.907840050  1.40510920
#> 46  046  0.600281780  0.69373179
#> 47  047 -0.241660679 -0.32200993
#> 48  048  0.169416927 -0.85947172
#> 49  049 -0.899470388 -1.05158186
#> 50  050 -0.113766046 -0.28284074
#> 51  051  1.144975682  0.51065058
#> 52  052 -0.279914038  0.60828267
#> 53  053  0.115691395  0.54568331
#> 54  054 -0.064707573  2.03899251
#> 55  055  0.231197382 -0.44157483
#> 56  056  1.385088397  0.89236497
#> 57  057  0.091371564 -0.64417384
#> 58  058 -1.446566114  1.39627954
#> 59  059 -0.509479796 -1.44566233
#> 60  060  0.507248200  0.47136310
#> 61  061 -0.923771300 -0.13740798
#> 62  062  0.674022910 -0.59300687
#> 63  063 -0.684052897 -1.76255620
#> 64  064 -1.036034707 -0.30280811
#> 65  065  0.978518935  0.99799405
#> 66  066  1.196638077  0.02496175
#> 67  067  0.854911075 -0.19095003
#> 68  068 -0.807436740 -0.66672354
#> 69  069  0.208470392  0.07293824
#> 70  070 -0.653750450 -0.48998875
#> 71  071 -0.804565956  0.31693736
#> 72  072  0.662635757  1.60064077
#> 73  073  0.569555280  0.33796612
#> 74  074 -1.150313835  0.26194816
#> 75  075  0.475105759  0.54791761
#> 76  076  0.373026441  0.85552437
#> 77  077  0.006170028  0.89204013
#> 78  078 -0.158617440  0.36866150
#> 79  079 -1.165932805 -1.10845797
#> 80  080 -0.301180304 -0.02438774
#> 81  081 -0.182937451  1.20420192
#> 82  082 -1.869178546 -1.13087299
#> 83  083  1.706838738  0.62279504
#> 84  084 -0.034453238 -1.01012060
#> 85  085 -0.141292035  1.58533050
#> 86  086  1.285187538 -0.37946410
#> 87  087  0.202706035  1.17726019
#> 88  088  1.952166655  1.79065834
#> 89  089 -1.459800502  0.78054602
#> 90  090  1.387500775  0.68183191
#> 91  091  0.651798243  0.53289610
#> 92  092 -0.034763862  0.18957275
#> 93  093 -0.160265363  0.15276818
#> 94  094  0.850830574  1.30696948
#> 95  095  0.053138644  0.58081422
#> 96  096 -0.197524643  0.95651519
#> 97  097 -0.469298898  0.71695015
#> 98  098  1.273631246  1.21715081
#> 99  099 -0.850264004 -1.01311121
#> 100 100 -1.254167545 -2.15115157