Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5182143
#> Y_2 0.5182143 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1          Y_2
#> 1 -0.1679250 -0.076631115
#> 2  0.3758509  0.008232673
#> 3 -1.6308006  1.972969283
#> 4  0.6365888  0.579668901
#> 5  1.2962448  1.558608098

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  2.153006291  0.621101719
#> 2   002 -1.526649060 -0.694111715
#> 3   003  1.360993658  2.622543391
#> 4   004 -1.398498639 -0.454059201
#> 5   005 -1.514164092 -1.366522874
#> 6   006 -0.223050148 -1.953013580
#> 7   007  0.983185826  0.962182899
#> 8   008  0.411626349 -0.094608753
#> 9   009 -1.204162782  0.851853090
#> 10  010  1.866298752  0.234002335
#> 11  011 -0.509759766 -0.811859808
#> 12  012 -0.077833466  2.235272764
#> 13  013  0.319326502  0.016969083
#> 14  014 -0.404630407 -0.298622463
#> 15  015  1.746553908  0.692913603
#> 16  016  0.205487061 -0.567703713
#> 17  017 -0.177454065  0.190111227
#> 18  018  0.759899817  0.264477209
#> 19  019 -2.067803596 -1.557770791
#> 20  020  0.097581716  0.325533761
#> 21  021 -0.393735674  0.917282344
#> 22  022  1.246366010  1.403386284
#> 23  023 -1.758092441 -0.433320671
#> 24  024 -1.245696023 -1.023616341
#> 25  025  0.088183014 -1.598591494
#> 26  026 -0.180747915  0.812181790
#> 27  027  1.364102896  1.988651365
#> 28  028 -1.253692064 -2.113334892
#> 29  029  0.833849069  0.433786873
#> 30  030  0.135368809  1.538767816
#> 31  031  0.731573538  0.232664523
#> 32  032 -0.009116732 -0.970682944
#> 33  033 -2.712511923 -0.785685840
#> 34  034 -0.260769035  0.800084472
#> 35  035 -0.966443067  0.420303301
#> 36  036 -0.062794108 -2.302563334
#> 37  037  0.461621034  0.909328776
#> 38  038  1.181777266  1.234324332
#> 39  039 -1.787309359 -2.163477259
#> 40  040 -0.873346373 -0.344131934
#> 41  041  0.672180673 -1.937235876
#> 42  042  0.185826851 -0.115786684
#> 43  043  0.992906153  0.816081310
#> 44  044  0.698797030  1.544937083
#> 45  045 -0.002279990 -0.793055312
#> 46  046 -1.529986719  0.201712741
#> 47  047  1.488172402  0.745781889
#> 48  048 -0.905497910  0.160097925
#> 49  049  2.009641405  0.845542605
#> 50  050  0.747642633 -0.468869449
#> 51  051 -0.487552278 -1.005505506
#> 52  052  0.027105424  0.422501453
#> 53  053 -0.993931950 -1.103709761
#> 54  054 -1.762484482 -1.636738009
#> 55  055 -0.697203970 -0.242587666
#> 56  056  0.928008265  1.803961458
#> 57  057  2.502225738  1.847432472
#> 58  058 -0.181289957 -0.150653155
#> 59  059 -0.664551317 -1.108160442
#> 60  060  0.703333216 -1.042718736
#> 61  061  1.416705229  0.091220016
#> 62  062  0.006241375  0.141135971
#> 63  063  0.758950016  1.694763105
#> 64  064 -0.334080575  0.304973486
#> 65  065 -0.162412538 -1.625660120
#> 66  066  0.286972677  0.017825430
#> 67  067 -1.792245383 -0.229877325
#> 68  068  1.143222845  0.003258099
#> 69  069 -1.455768639 -2.628499987
#> 70  070 -0.139354491 -0.674750630
#> 71  071  0.097944769 -0.904443707
#> 72  072 -0.957166527  0.006268782
#> 73  073 -0.204894313  1.438135599
#> 74  074  1.824926411  2.837164824
#> 75  075  0.022947052  0.242031708
#> 76  076 -0.803177127 -1.235107115
#> 77  077  0.109486324  0.343734428
#> 78  078 -1.496138613  0.729107039
#> 79  079 -0.897145565 -1.453278715
#> 80  080 -0.624557642 -1.396379151
#> 81  081  0.449202165 -0.578975543
#> 82  082 -0.181793588  0.076194334
#> 83  083 -0.171844142  0.102447026
#> 84  084  0.999763511  0.148390140
#> 85  085 -3.037548700 -1.047098621
#> 86  086 -0.023169139 -0.482553110
#> 87  087  0.904864014 -2.039265851
#> 88  088 -0.702152016 -0.678817281
#> 89  089 -0.178190284  0.158847659
#> 90  090 -1.321843378  0.800137381
#> 91  091 -0.893251805  0.790639910
#> 92  092  1.031271716  0.244389703
#> 93  093 -1.642909108 -0.639720120
#> 94  094  1.334008229  0.459032619
#> 95  095  0.425840673  0.660327015
#> 96  096  0.061592452 -1.829714074
#> 97  097 -0.941816373 -1.711359112
#> 98  098  0.988197262  1.183486690
#> 99  099 -1.135143236 -0.896354136
#> 100 100  0.466761745 -2.073669728

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  0.4016720823  0.22833940
#> 2   002 -0.6900538126  0.36762340
#> 3   003 -0.3779828065 -1.32739104
#> 4   004  0.6338203800 -0.33600459
#> 5   005  1.0616136510  1.94102028
#> 6   006  0.2661766208 -0.56058637
#> 7   007  0.3203687517 -0.06584859
#> 8   008 -0.8728686433 -1.06984211
#> 9   009 -0.4276348292 -0.76470768
#> 10  010  1.2577598229 -0.35046536
#> 11  011  0.6175222968 -0.59135219
#> 12  012 -2.8896798344 -0.38007818
#> 13  013  0.9883570836 -1.12209590
#> 14  014 -0.2400559689  0.97260584
#> 15  015  0.5608006951 -0.01422375
#> 16  016 -1.3182806024 -2.08337827
#> 17  017 -1.5533622084 -0.33568972
#> 18  018  0.7307645208  0.25206189
#> 19  019 -0.1031348097  0.69239769
#> 20  020 -1.7583123149 -1.90467974
#> 21  021 -0.2780073873  0.23757697
#> 22  022 -0.0846178656 -0.70524321
#> 23  023 -0.3021268479  1.35134622
#> 24  024 -0.5614299033 -0.91957629
#> 25  025 -0.8248391805 -0.16907925
#> 26  026  1.9749798587  1.18943709
#> 27  027  0.2646345349  0.21488502
#> 28  028  3.1303491217  1.65015328
#> 29  029 -0.4231330488  1.61961657
#> 30  030 -0.5114666736  1.21744324
#> 31  031 -0.9617569625  0.45605140
#> 32  032  0.9216954984  0.67634949
#> 33  033  0.3983607510 -0.89342202
#> 34  034  0.4613927627 -0.04992695
#> 35  035  0.2199650981 -0.37010465
#> 36  036 -1.2979415982 -1.89814387
#> 37  037 -0.1962129633 -0.11802724
#> 38  038 -0.9348338824 -2.14987664
#> 39  039  0.3944132148  0.68084253
#> 40  040 -0.9091448600  0.17420971
#> 41  041  0.6718027233  0.51092677
#> 42  042  1.4302514625  0.24433529
#> 43  043  0.6189222328  0.38696326
#> 44  044 -0.2037382375  1.23203049
#> 45  045  0.5961879203  2.02499581
#> 46  046  0.1514389315 -0.43868218
#> 47  047  0.1327677373  0.83866102
#> 48  048 -1.1997068908 -0.63729386
#> 49  049 -0.2890724038  1.71286807
#> 50  050  0.4790327094  0.40292730
#> 51  051 -0.6241313753 -0.05956114
#> 52  052  0.1023938557  1.77092777
#> 53  053  1.4288841793  1.15270718
#> 54  054  0.1154950674 -0.78442609
#> 55  055  0.0908891084 -0.64960115
#> 56  056  0.8812506283 -0.16899066
#> 57  057  0.9503253867  0.70465340
#> 58  058  1.0779022251  1.45009518
#> 59  059 -0.1298934919  0.28008426
#> 60  060 -0.9928274191 -1.41941086
#> 61  061  0.8717569977  2.67854256
#> 62  062  1.5938487878  0.62786059
#> 63  063 -0.4752317512  0.26080251
#> 64  064 -0.4248994044 -0.75785439
#> 65  065 -1.0855326897  0.68762338
#> 66  066  0.7921102222  2.87692212
#> 67  067  1.4415100555  0.51178617
#> 68  068 -1.7657202244 -0.50821840
#> 69  069 -0.5404821351 -1.05460763
#> 70  070  1.3588160679  1.11862178
#> 71  071 -0.5950549826  0.22021665
#> 72  072  1.8439795622 -1.01969342
#> 73  073 -1.3742660604 -2.17828321
#> 74  074  2.1155288925  0.94736980
#> 75  075  0.3521620299 -0.48005452
#> 76  076  0.4394404526 -0.16823324
#> 77  077  0.9270106169  0.50747510
#> 78  078 -0.8736579161  0.55342187
#> 79  079  0.3323921282  1.98294926
#> 80  080  0.0006364369  0.62414710
#> 81  081  0.8661974954  0.73424503
#> 82  082  0.6945580527  0.24953813
#> 83  083 -0.6333710650  0.07548014
#> 84  084 -0.1660863881  0.21798825
#> 85  085  1.2693732349  0.97463986
#> 86  086 -0.4314498195 -1.13621266
#> 87  087 -1.5232767620 -0.86187184
#> 88  088  0.4751444784  0.40210479
#> 89  089  0.1322767730  0.92000852
#> 90  090 -1.8285021915 -1.05885866
#> 91  091 -0.4732434699 -0.01164494
#> 92  092  1.7766752482  0.67541062
#> 93  093  0.6416402203 -1.64720164
#> 94  094 -1.6237183257 -0.53543981
#> 95  095 -2.0080350740 -0.49166453
#> 96  096 -1.3547493056 -0.49957748
#> 97  097  1.5687373418  0.85503197
#> 98  098  1.0570114333  0.45729430
#> 99  099 -0.7209753038 -0.37794809
#> 100 100  0.3781882751 -0.56991419