Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.510421
#> Y_2 0.510421 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.2886244  0.3451977
#> 2 -0.1568524  0.2190844
#> 3 -0.3050897  0.2398503
#> 4 -0.3982981 -1.3191974
#> 5  0.6787286  1.5115104

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001  2.853039164  2.60240911
#> 2   002  0.680091042 -0.46970739
#> 3   003 -0.501818079 -1.72443534
#> 4   004 -1.729392375 -1.00130779
#> 5   005  1.656579115  1.91693843
#> 6   006 -1.274614070 -0.02755869
#> 7   007 -0.224026060  1.30630916
#> 8   008 -1.093527861 -1.14633210
#> 9   009  1.851794177  1.11808642
#> 10  010 -1.942202355 -1.57265018
#> 11  011  1.255599244  0.29635590
#> 12  012 -1.321666428 -0.12938479
#> 13  013  0.717185808  0.77760515
#> 14  014 -1.508363929 -1.11951785
#> 15  015  0.509162250  1.25363666
#> 16  016 -0.791777384  1.17685148
#> 17  017  0.811406356  0.54989174
#> 18  018 -0.553073038  2.24443589
#> 19  019  0.498033989  1.29999471
#> 20  020  1.172687068  0.63888896
#> 21  021  0.471920271  1.00230318
#> 22  022 -0.740244326 -0.24445751
#> 23  023  2.259987801  2.65066252
#> 24  024  1.219320319  0.11595261
#> 25  025  0.796830117  0.07589471
#> 26  026 -1.495247174 -0.61671389
#> 27  027  1.964310397 -0.31638229
#> 28  028  0.410020887 -0.13667744
#> 29  029  0.409117335  0.12471719
#> 30  030  0.033220409 -0.57867624
#> 31  031 -1.927997215 -1.35515425
#> 32  032  1.913958971  1.51788737
#> 33  033  0.396419783 -0.52959165
#> 34  034  0.026542079 -0.84842119
#> 35  035 -0.830757279 -0.88128148
#> 36  036  1.723121473  1.57536455
#> 37  037  0.098721831 -0.90342081
#> 38  038 -0.478380528 -0.49517232
#> 39  039  1.222053910  0.72997249
#> 40  040 -0.837271990 -0.39962287
#> 41  041 -0.010941126 -0.76828066
#> 42  042  0.383831886 -0.77700463
#> 43  043  2.154167036  1.32193683
#> 44  044 -0.860923367  0.99190389
#> 45  045  0.493461970 -0.38151406
#> 46  046 -0.407509367 -0.49462418
#> 47  047 -0.166046366 -0.02373104
#> 48  048 -1.044062862 -0.77061195
#> 49  049 -0.604108502 -0.94724024
#> 50  050  0.899339676  0.67630394
#> 51  051 -0.008441077  1.10493141
#> 52  052 -0.176413308  0.63462134
#> 53  053 -3.446191462 -1.39313870
#> 54  054 -0.593550291 -0.12929429
#> 55  055 -1.069454075  0.90284741
#> 56  056 -1.587637616  0.11279905
#> 57  057  1.742692905  0.18408313
#> 58  058 -0.099689807 -0.59179827
#> 59  059 -1.257096468 -2.79076124
#> 60  060  0.771752111  1.05805927
#> 61  061 -1.287538905 -1.04389417
#> 62  062 -0.116052891  1.71723655
#> 63  063 -0.375557248 -0.06567564
#> 64  064 -0.764840825 -0.24855124
#> 65  065  1.395208342  1.23177293
#> 66  066  0.526205412 -0.36914044
#> 67  067 -1.262986681 -0.44649822
#> 68  068 -1.384182629 -0.85021217
#> 69  069  0.781593073 -0.99226674
#> 70  070  1.262053848  0.04705866
#> 71  071  1.261801323  1.06154308
#> 72  072 -1.387307752 -0.55699776
#> 73  073  0.140463961 -1.07509044
#> 74  074 -0.783219893  0.14806055
#> 75  075  1.057341390 -0.45186389
#> 76  076  0.282272457 -0.28824580
#> 77  077  0.967334145  0.11472435
#> 78  078 -0.428148038 -0.29847994
#> 79  079 -0.663431172 -2.48760006
#> 80  080  0.458161287  0.37991589
#> 81  081  0.426666275 -0.32906092
#> 82  082 -0.650647184  0.56168073
#> 83  083 -0.824594802  0.04685078
#> 84  084  1.478553642 -0.11458487
#> 85  085  0.423882198  0.26738372
#> 86  086  0.852011385  2.33784098
#> 87  087  0.886409804 -1.22299761
#> 88  088 -0.946644280 -1.78982130
#> 89  089  0.697569903  0.32729079
#> 90  090 -0.369949794 -0.91650571
#> 91  091 -0.581804203  0.24154344
#> 92  092 -2.325957014 -2.27060203
#> 93  093 -0.162486032 -1.60174105
#> 94  094 -0.270421860 -0.18193964
#> 95  095 -0.651595117 -0.08639503
#> 96  096  2.415093960  1.14445412
#> 97  097  0.163930179  0.19550837
#> 98  098  0.076043413 -1.00318561
#> 99  099 -1.410874192 -0.40234044
#> 100 100  0.219867718  0.01055059

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.59515197  0.077025184
#> 2   002  0.99326773  0.796893334
#> 3   003  0.49513941  2.677633503
#> 4   004 -0.32320587  1.768873590
#> 5   005  1.72394457  0.665482918
#> 6   006  0.53117860 -0.565681498
#> 7   007  0.67814003 -0.596637728
#> 8   008  1.43708377  1.622686096
#> 9   009 -1.76016282 -1.380865309
#> 10  010  0.94510811 -0.799513995
#> 11  011 -0.65498051 -0.887941922
#> 12  012 -0.20122007 -0.613072582
#> 13  013  1.67099108  1.579237107
#> 14  014  0.07440181 -1.682971021
#> 15  015 -0.21804188  0.732414416
#> 16  016 -0.87922375 -0.571819346
#> 17  017  0.32119493  0.524073093
#> 18  018  0.85413588  0.197551689
#> 19  019 -1.10191976  0.045887799
#> 20  020  0.98734722  1.085403202
#> 21  021  0.54460716  0.113590700
#> 22  022  0.41804986 -0.279548784
#> 23  023 -0.33553536  0.419615652
#> 24  024 -0.56406390  0.286646184
#> 25  025  0.45103412  0.198278347
#> 26  026  0.69983936  0.577917820
#> 27  027  1.00393681  1.578950990
#> 28  028  0.13217946  0.398398248
#> 29  029 -0.64391838 -0.189138907
#> 30  030  0.26549251  0.722469966
#> 31  031 -0.48901389  0.006171756
#> 32  032 -0.54928086 -1.461419118
#> 33  033  0.89022214  0.508823404
#> 34  034 -0.34020842 -0.082154520
#> 35  035  0.73096930  0.831636655
#> 36  036 -0.43856273 -0.444926916
#> 37  037 -1.34870294 -0.424171652
#> 38  038 -0.68838241 -1.557142386
#> 39  039  0.01315390 -0.045197681
#> 40  040  0.14278599 -0.597884127
#> 41  041  1.59807562  1.491222595
#> 42  042  0.93092793 -0.328446031
#> 43  043  0.31931681 -0.110889527
#> 44  044 -1.23753685 -1.152218452
#> 45  045  0.64719277 -0.337193420
#> 46  046  1.41656517  0.065621752
#> 47  047 -1.13847717 -0.372268696
#> 48  048  1.02501193  0.007003082
#> 49  049  0.69187631 -0.421941469
#> 50  050 -0.09054958 -0.658651794
#> 51  051 -0.98706264 -0.601684139
#> 52  052 -0.02489369 -1.006183871
#> 53  053  0.34582936  0.064892508
#> 54  054  0.86271956 -0.566463044
#> 55  055 -0.38877668 -0.991263797
#> 56  056  0.70234285 -0.357437065
#> 57  057  1.26782697 -0.117868293
#> 58  058 -0.14137087 -0.254277758
#> 59  059  0.33912690  1.529853666
#> 60  060 -0.21977309  0.817962992
#> 61  061 -0.23552874  0.498938884
#> 62  062  0.34488996  0.717505307
#> 63  063 -0.45232192 -0.541484146
#> 64  064 -1.02899488  0.141732413
#> 65  065 -0.59454873 -0.441772389
#> 66  066 -2.87363243 -2.180889824
#> 67  067  0.88736734  2.587395520
#> 68  068  0.93793203 -0.461613446
#> 69  069 -0.03662406 -1.911481836
#> 70  070  0.61545210 -0.468436449
#> 71  071  0.22475374 -0.151581911
#> 72  072  0.87439736  1.394943612
#> 73  073  0.49524969  0.891269938
#> 74  074  2.40056760  2.287304446
#> 75  075 -0.31484506 -1.385913817
#> 76  076 -1.59577945 -2.179738687
#> 77  077 -0.77434493 -0.962604477
#> 78  078 -0.05125633  0.926913988
#> 79  079 -1.81330579 -0.466778847
#> 80  080 -0.04333106 -0.323613039
#> 81  081  0.09057424  0.208678018
#> 82  082 -0.75863807 -1.482042883
#> 83  083 -1.58207756 -1.091918357
#> 84  084 -1.12026095 -0.890220991
#> 85  085 -1.36509314 -0.632054997
#> 86  086  0.65452848  0.497414095
#> 87  087  1.79646778  0.428315071
#> 88  088 -0.59080784  0.190080572
#> 89  089 -1.52160502  0.113045623
#> 90  090 -1.34497449  0.600492914
#> 91  091  0.50215234  0.508778823
#> 92  092  0.72561481 -0.713294348
#> 93  093  0.40513060  0.400138378
#> 94  094  0.88949543  0.335981894
#> 95  095  1.53938101  0.773630733
#> 96  096 -2.06896489  0.447508733
#> 97  097  1.34795458  0.064609773
#> 98  098  0.33591952  1.584405653
#> 99  099 -0.48083003  0.033418421
#> 100 100 -0.13992317  0.516031020