Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4726893
#> Y_2 0.4726893 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1278198  0.1027314
#> 2 -0.1628628  0.2475787
#> 3 -0.3559527 -0.7890083
#> 4  0.1906511  0.4217700
#> 5 -0.9332282 -1.1063808

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.92739524  0.205826926
#> 2   002 -0.43214923 -1.098676504
#> 3   003  0.52874687 -0.689687966
#> 4   004 -0.60433292 -0.614452026
#> 5   005  0.45202365 -1.003390323
#> 6   006 -0.76604424 -0.821961749
#> 7   007  1.09444813  0.269865959
#> 8   008  0.61431716 -0.133541657
#> 9   009  1.04029026  0.079954017
#> 10  010 -0.30672635  1.645824493
#> 11  011  1.50215853 -0.934548373
#> 12  012 -0.37410085  0.217420059
#> 13  013  0.33029506  0.932081053
#> 14  014 -2.25879160 -2.852512252
#> 15  015 -0.65401503 -0.587897690
#> 16  016  2.75172892  1.063081016
#> 17  017 -0.99996651  1.113266795
#> 18  018 -0.85791032 -1.721664843
#> 19  019  0.55692965  0.539881423
#> 20  020  1.00814074  0.248032675
#> 21  021 -0.85363356 -0.215463454
#> 22  022 -0.17187250 -0.626255786
#> 23  023 -0.22790010 -0.363753240
#> 24  024 -0.75991038 -0.638887112
#> 25  025 -0.69306807 -0.748492973
#> 26  026  0.63847237  0.196352230
#> 27  027 -1.23680942 -0.174599715
#> 28  028 -0.14472005 -1.526099410
#> 29  029  0.58678870 -0.309544173
#> 30  030  0.86939065 -0.242582880
#> 31  031  0.17935789  0.472875395
#> 32  032 -0.03680483 -0.690394059
#> 33  033  0.62419408  1.724963156
#> 34  034 -1.16800804  0.046909947
#> 35  035  1.49338390 -0.638685133
#> 36  036  0.35648172 -0.849718411
#> 37  037  0.01057487 -0.853907474
#> 38  038 -0.18669678 -1.595035115
#> 39  039 -0.12850629  1.177438482
#> 40  040 -0.69068035 -0.009136942
#> 41  041  1.35948805  0.385900780
#> 42  042 -1.02528404  0.695516653
#> 43  043  1.64594251  1.328957614
#> 44  044  1.00242314 -0.834079675
#> 45  045  1.47899939  1.067489045
#> 46  046 -0.92733543 -0.863820489
#> 47  047 -0.96706896 -0.221833956
#> 48  048 -1.20859239  0.330089277
#> 49  049 -0.51765091 -0.201746134
#> 50  050 -1.95560523 -1.114524153
#> 51  051  0.49883153 -1.271294728
#> 52  052 -1.22394896 -0.241836072
#> 53  053  0.74527363  1.051057119
#> 54  054  0.18024985  0.420253526
#> 55  055 -0.62405425 -0.206049070
#> 56  056 -0.79096840 -0.360131530
#> 57  057 -1.18249920 -0.859470969
#> 58  058  1.98447728  2.716085832
#> 59  059  0.15326279 -0.874664498
#> 60  060  1.29740307  0.039192980
#> 61  061 -0.87060153 -1.275837507
#> 62  062 -0.50803796  0.723197133
#> 63  063  0.87795927  0.857003400
#> 64  064  0.68130850  1.053247050
#> 65  065  0.05809259 -0.998459733
#> 66  066 -0.77372020 -0.594433432
#> 67  067  0.36878980  0.282326723
#> 68  068  0.55328775  0.528409122
#> 69  069 -1.28798548 -2.505773565
#> 70  070 -1.62368438 -1.633060779
#> 71  071  1.37283465  1.892326635
#> 72  072  0.56769861  1.920899940
#> 73  073 -0.24770758  0.295571021
#> 74  074  0.68417881  0.743372305
#> 75  075 -1.48082167 -0.230192316
#> 76  076  1.00936136 -0.397184030
#> 77  077 -1.56541111 -0.410971926
#> 78  078  0.92230060  0.631187824
#> 79  079  0.43432496  0.620695755
#> 80  080 -0.20880816  1.042259460
#> 81  081  2.50889913  0.264050840
#> 82  082 -1.32808117  0.338695915
#> 83  083  1.47106817 -1.285756799
#> 84  084 -0.83295538 -1.313701581
#> 85  085 -0.29070198  0.073896887
#> 86  086 -0.83922230  0.527802975
#> 87  087  1.10522824  0.558656774
#> 88  088 -0.05245224 -1.343083947
#> 89  089  0.14611053 -0.482593430
#> 90  090 -0.24573335  1.162161553
#> 91  091 -1.73401994 -0.067073209
#> 92  092 -0.85309125 -3.185853449
#> 93  093  1.23416402  1.067915192
#> 94  094  1.27591438 -0.421644826
#> 95  095 -0.53524189 -0.748746831
#> 96  096  1.55436306  0.004556830
#> 97  097 -0.14380943  0.036795257
#> 98  098 -1.70455621  0.299851605
#> 99  099 -1.72257341  0.344142195
#> 100 100 -1.56355100 -1.218031725

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.555118033 -0.19128224
#> 2   002  0.506817710 -0.41984992
#> 3   003 -0.539730664 -0.06137229
#> 4   004  1.146355346 -0.57561798
#> 5   005 -0.078595915 -1.18337501
#> 6   006  0.546713815  1.05810303
#> 7   007 -0.098948029  0.17823698
#> 8   008  1.034161191 -0.80327868
#> 9   009  0.928573375  1.00093087
#> 10  010 -1.910151612  0.67281492
#> 11  011  0.714492263  1.26499222
#> 12  012  1.578600939  1.13827180
#> 13  013 -0.163843320  1.23610298
#> 14  014  0.533358258 -1.65692988
#> 15  015  0.036328258  1.17445092
#> 16  016 -0.559198327 -1.22145651
#> 17  017  0.400758684  0.57788698
#> 18  018 -0.317202704  0.36082237
#> 19  019  1.513758232 -0.78965807
#> 20  020  0.446446259  1.19976850
#> 21  021 -0.187780600 -0.25291286
#> 22  022  2.635874924 -0.07417081
#> 23  023 -1.957158648 -0.85847988
#> 24  024 -0.447137692 -1.01680739
#> 25  025 -0.322533712  0.69552737
#> 26  026  1.368279019  1.20822890
#> 27  027 -0.912091391 -0.56705469
#> 28  028  1.893364576  0.37396435
#> 29  029 -0.777366232 -0.33494511
#> 30  030  0.200727926 -0.81704065
#> 31  031 -1.056059261 -1.78013175
#> 32  032  0.756349506  2.09026416
#> 33  033  0.435160669  1.69155768
#> 34  034  0.439143958  2.27033647
#> 35  035 -1.353187154 -0.60826009
#> 36  036 -1.769317128 -0.25132228
#> 37  037 -0.763390773  0.02991291
#> 38  038  2.230960068  1.45491543
#> 39  039  1.149989237  0.37954990
#> 40  040 -0.986121211 -0.46553108
#> 41  041  0.480511367  0.66070380
#> 42  042  0.631330182  1.24686057
#> 43  043  1.425684772  1.03833568
#> 44  044  0.968431790  0.28183647
#> 45  045 -0.619837608 -1.99557727
#> 46  046  0.438420446  1.39509060
#> 47  047 -0.082436549  0.15523944
#> 48  048  0.901563262  1.02882115
#> 49  049  1.032779963  0.24567290
#> 50  050  1.248998324 -0.52836540
#> 51  051 -1.438481421 -1.14397881
#> 52  052  0.819506489 -0.32601990
#> 53  053 -0.603066770 -1.07465344
#> 54  054  0.125802178  0.78928057
#> 55  055  0.177397103  0.84492009
#> 56  056  0.404040694  0.77072791
#> 57  057 -1.256332096  0.02090416
#> 58  058 -0.435458645 -0.13520310
#> 59  059 -2.035182519 -0.36820474
#> 60  060 -0.144420912 -1.29937106
#> 61  061 -1.084963451 -1.17769444
#> 62  062  0.352967983 -2.76174641
#> 63  063  1.483483563  0.07642699
#> 64  064  0.717974669  0.72838768
#> 65  065  0.934439227  1.33411678
#> 66  066  1.511933360 -0.77775336
#> 67  067  0.604454083  0.19310222
#> 68  068  0.263814208 -0.63170138
#> 69  069 -0.245721825 -2.62617083
#> 70  070 -0.773708212  0.11936959
#> 71  071 -0.993930559 -1.67996453
#> 72  072  0.755329303 -0.06811335
#> 73  073  0.891946381  1.38585214
#> 74  074  0.006338242 -0.49690040
#> 75  075  0.446976889  0.05308201
#> 76  076 -0.639498672 -0.87202306
#> 77  077  0.147623811 -0.46507592
#> 78  078  0.159424748  0.89491750
#> 79  079 -0.895954558 -1.48959751
#> 80  080  0.334715170 -0.15895200
#> 81  081 -0.425072640  0.37147169
#> 82  082  0.080831542 -1.14210920
#> 83  083  0.016754941  0.51972110
#> 84  084  0.482219508  1.10883778
#> 85  085  0.639265763 -0.05956415
#> 86  086  0.404768056  0.59266014
#> 87  087 -0.706671897 -0.53648160
#> 88  088 -0.224938898  0.76059734
#> 89  089  0.187025075  1.07974061
#> 90  090  0.626937936  0.99862304
#> 91  091 -2.559794770 -1.41148426
#> 92  092 -0.680480751 -0.40560539
#> 93  093 -1.244453091 -0.81323896
#> 94  094 -1.133613050 -0.01916716
#> 95  095  0.675875642  1.03948291
#> 96  096 -1.524439235 -1.51007404
#> 97  097 -0.599760313  0.01892782
#> 98  098 -0.234366481  0.20339816
#> 99  099 -0.096811155 -1.25892257
#> 100 100  0.328390951  0.04628357