Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5245135
#> Y_2 0.5245135 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.4263145  0.04244247
#> 2 -0.3293053  0.28220389
#> 3 -0.7210446 -0.56770355
#> 4  0.5067209  1.10157798
#> 5 -1.2951216 -0.70073388

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.78817351 -0.42913163
#> 2   002  1.79749604  1.55289635
#> 3   003 -0.81782294 -1.11712513
#> 4   004 -0.61493224  0.11999004
#> 5   005  0.24623269 -0.14017763
#> 6   006  1.88313081  0.32043815
#> 7   007 -0.57251647  0.93691080
#> 8   008 -0.63099675  0.83331001
#> 9   009 -0.39602760 -1.47830189
#> 10  010  0.39883455 -0.39960573
#> 11  011  1.26650054  1.41306115
#> 12  012 -0.29909125  0.50043637
#> 13  013  0.49179993  0.44662545
#> 14  014  0.72705228 -0.87646611
#> 15  015 -0.98492437 -1.28243278
#> 16  016  1.78485310  1.86167901
#> 17  017 -1.37408856  0.83060190
#> 18  018  1.96607627  0.87101069
#> 19  019 -0.95282563 -1.08412672
#> 20  020 -1.58015876 -1.66481673
#> 21  021 -0.68853827  0.53791445
#> 22  022 -1.63670172 -0.70988767
#> 23  023  1.56042881  1.43392229
#> 24  024  1.40522751 -0.25925443
#> 25  025  0.64180037  1.01672156
#> 26  026  0.76731036  0.96756206
#> 27  027 -0.86160986 -1.02853368
#> 28  028  0.87242245 -0.49336758
#> 29  029  0.20753142  1.36722695
#> 30  030  0.94560604  0.50849352
#> 31  031 -0.45788964  0.71498509
#> 32  032  0.51968683  0.25698281
#> 33  033  1.52338693 -1.53461177
#> 34  034 -0.39939889 -0.36284892
#> 35  035 -0.84800493 -2.09179676
#> 36  036 -0.72871275 -0.53056541
#> 37  037 -0.49595267 -0.23210237
#> 38  038  0.29605820  0.20839187
#> 39  039  0.40086873  2.24378244
#> 40  040 -0.08651410  0.99059378
#> 41  041 -0.27205137  1.37926504
#> 42  042  0.74833421  0.98262420
#> 43  043 -0.49197969 -2.30396071
#> 44  044  0.47472715 -0.22050516
#> 45  045  0.96257482  0.92902895
#> 46  046  0.72186884  0.62633667
#> 47  047 -2.00927969 -0.73797149
#> 48  048 -0.50362758 -1.46250481
#> 49  049 -0.20009615 -0.67408532
#> 50  050  1.41245326  0.78662104
#> 51  051  0.05572960  0.18354883
#> 52  052  2.59201514  2.14602269
#> 53  053 -1.05713935 -0.25876127
#> 54  054  1.90648738 -1.41754502
#> 55  055 -0.90271609 -0.13023696
#> 56  056  0.66051519  0.77216169
#> 57  057 -1.42688541 -1.18157921
#> 58  058  0.20386075 -0.84494778
#> 59  059  0.08442660 -2.88715103
#> 60  060 -1.84194691 -1.94946091
#> 61  061  0.04534498  0.90084962
#> 62  062  0.30917123  2.24430058
#> 63  063  1.84466983  1.63272531
#> 64  064 -0.74295121 -1.23818538
#> 65  065  1.08981843  1.21934958
#> 66  066  0.07939035  0.45865482
#> 67  067  0.88147121  0.60578984
#> 68  068  0.78790472 -1.37395532
#> 69  069  1.17069794  0.52721327
#> 70  070 -0.57821323 -0.83207912
#> 71  071  0.05795973 -0.38334919
#> 72  072  0.24246019 -0.05619871
#> 73  073  1.61454273 -0.10594999
#> 74  074 -0.11809519  1.44716595
#> 75  075  0.21181488  0.16331139
#> 76  076 -0.86123493 -0.17524469
#> 77  077 -1.12389146  0.14181205
#> 78  078  0.25838482 -0.13730990
#> 79  079 -1.62181586 -1.06459974
#> 80  080  0.29662774 -0.32620679
#> 81  081 -1.66349641  0.29979937
#> 82  082 -1.98819345  0.45827796
#> 83  083 -0.49408332  0.25911040
#> 84  084 -0.95472807 -2.68455968
#> 85  085 -2.74626699 -0.13549455
#> 86  086  0.49295223 -2.09995523
#> 87  087 -0.36924850 -1.16428731
#> 88  088  0.39572929  0.45593077
#> 89  089  0.50285431 -0.52357357
#> 90  090  0.12347527  0.84923011
#> 91  091 -0.05753766  0.28977775
#> 92  092  0.93229190 -0.20437026
#> 93  093 -1.34868544 -0.83571398
#> 94  094 -0.46731945 -1.33310565
#> 95  095 -2.05430052 -0.20818202
#> 96  096 -0.05281662  0.42676985
#> 97  097  2.13190305  1.26077497
#> 98  098 -0.80443286  1.10281294
#> 99  099  1.32911453  1.72687035
#> 100 100 -0.80890195 -0.04115566

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.101759523 -0.83677072
#> 2   002 -0.275360501 -0.45279572
#> 3   003 -0.600043772 -1.10719852
#> 4   004  0.191682864 -0.49399689
#> 5   005 -0.224304598 -0.40535222
#> 6   006 -0.475622096  0.34113409
#> 7   007  1.913181197  0.14808531
#> 8   008  0.607783335  1.00441972
#> 9   009  0.587275459  0.17333254
#> 10  010 -0.697552020 -0.44288672
#> 11  011  0.350044496  0.19081220
#> 12  012  0.411845580  0.85845499
#> 13  013 -0.180004342 -0.05222080
#> 14  014 -0.344070162  0.42250121
#> 15  015  0.543023070  0.68706245
#> 16  016 -1.011116845  0.13618395
#> 17  017  0.687397302 -0.11599842
#> 18  018  0.385640487 -0.55542761
#> 19  019  1.161368818  0.29511872
#> 20  020 -0.729368334  1.61165461
#> 21  021 -0.019782329 -0.03213724
#> 22  022 -2.161784456  0.44202852
#> 23  023  0.383277965 -0.04171859
#> 24  024 -1.609714407 -1.94554475
#> 25  025 -0.322242410 -0.43495102
#> 26  026  0.119672409 -0.44841612
#> 27  027 -0.393538565  0.01663076
#> 28  028  0.382498809  0.40422121
#> 29  029  0.578004832  1.06261290
#> 30  030 -1.839582466 -1.04333509
#> 31  031 -0.354896961  1.03574921
#> 32  032  1.849908863  1.00730731
#> 33  033  0.175445354 -0.75738262
#> 34  034 -1.565049970  0.37823963
#> 35  035 -0.009735834 -0.16278239
#> 36  036 -0.076954796 -0.16541883
#> 37  037  0.655937438  0.19569111
#> 38  038 -0.108822436  1.48023215
#> 39  039  1.423106307  0.62903711
#> 40  040  1.380149362  0.71280674
#> 41  041  1.145878966  1.02510204
#> 42  042  0.529764437 -0.76429521
#> 43  043  0.812726291  1.64621692
#> 44  044  0.089755878  0.45686456
#> 45  045  0.566053856 -0.53578962
#> 46  046  0.871233589  0.07578794
#> 47  047 -2.321520909 -0.58545880
#> 48  048 -0.334853447 -0.40560605
#> 49  049 -0.168543083 -0.45152625
#> 50  050  0.541234036 -1.08555033
#> 51  051  0.994427277  2.75629864
#> 52  052 -1.092205811 -1.48275945
#> 53  053  0.533017921  0.14083337
#> 54  054  0.005643781 -0.14582161
#> 55  055 -0.700811771 -0.90675790
#> 56  056  1.558183518 -0.16649112
#> 57  057 -1.605746855 -0.81272990
#> 58  058  0.650993480 -0.11080999
#> 59  059 -0.513960992 -0.20940630
#> 60  060  0.606525370 -0.10463349
#> 61  061 -0.709787538 -2.18437812
#> 62  062 -1.109336443 -0.64408035
#> 63  063  1.635509327  0.83697446
#> 64  064 -0.937370643  0.70324077
#> 65  065  0.228043740  0.45013939
#> 66  066  0.200816576  0.45798871
#> 67  067 -1.150738263  0.86367921
#> 68  068 -0.191583588  1.55162254
#> 69  069 -0.368361982 -0.40783523
#> 70  070  1.409941440  0.21644418
#> 71  071 -0.396788326  0.31040892
#> 72  072  0.419519838 -0.20776184
#> 73  073  0.010300330  0.22785013
#> 74  074 -1.503423927  1.76908151
#> 75  075  1.635667664  1.88312817
#> 76  076 -0.935207472  0.01883837
#> 77  077 -0.383876592 -2.36281165
#> 78  078  0.084791347  0.02367145
#> 79  079 -0.985912180 -0.53117302
#> 80  080 -2.150238249 -0.06695855
#> 81  081  0.214779233 -0.51154281
#> 82  082  0.477230401 -0.07218250
#> 83  083  0.165150348  0.04210362
#> 84  084  0.270432812  1.00090479
#> 85  085  0.766984249  0.18578115
#> 86  086  0.909826223 -1.43698652
#> 87  087  0.716594427  1.80175907
#> 88  088 -0.646155405  0.03191486
#> 89  089  0.451674639 -0.23957569
#> 90  090 -1.273493681  0.80532189
#> 91  091 -2.421823381 -1.15396729
#> 92  092  0.875415543  0.72194260
#> 93  093  0.056941328  0.16251443
#> 94  094  0.252350515 -0.07742297
#> 95  095 -0.203332698  0.31072861
#> 96  096  0.978692902 -0.96651470
#> 97  097  0.071252356 -1.39301577
#> 98  098 -1.390382580 -1.09823268
#> 99  099 -1.356407130 -0.58500306
#> 100 100  0.427991125 -1.08600590