Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4077945
#> Y_2 0.4077945 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1  1.03409381  1.1392114
#> 2  1.29934503 -0.6681197
#> 3 -0.06043217  1.4804374
#> 4 -0.33875994 -0.2932946
#> 5  0.76263516  0.7043875

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1           Y_2
#> 1   001 -0.65806838 -1.5303993970
#> 2   002 -2.65234568 -1.3018450171
#> 3   003 -0.30406062  0.3925703373
#> 4   004 -0.14766801 -0.0438697769
#> 5   005 -0.49981205 -0.6070312867
#> 6   006  1.10524890  0.2726808730
#> 7   007  1.13026592  1.1598353310
#> 8   008 -1.43556894  0.7458729891
#> 9   009  0.15324390  2.6285391072
#> 10  010 -0.71881580  0.1817064656
#> 11  011  0.64019111  1.8671684848
#> 12  012  1.15200996  1.0462914629
#> 13  013  0.64891698  0.8909310455
#> 14  014  0.68616220  0.4370237692
#> 15  015 -2.15711206 -0.7717029911
#> 16  016  0.98480233 -0.0251110323
#> 17  017 -2.19855434 -1.4383888964
#> 18  018  1.07936613  1.6711879362
#> 19  019  1.11382504  0.4344353614
#> 20  020 -1.10784390 -0.7883519752
#> 21  021 -0.94899471  0.1841425401
#> 22  022  0.86497866  0.1859860002
#> 23  023  0.53766950  1.4860775370
#> 24  024  0.65157255  1.0890226384
#> 25  025  0.47446564  1.3440523211
#> 26  026  1.43977644  1.7517293537
#> 27  027 -0.41192455 -0.1175935630
#> 28  028  1.08409590 -0.7611145968
#> 29  029  0.93731134  1.2356544913
#> 30  030  0.22873368 -0.2285424310
#> 31  031  0.90426993  0.5448173072
#> 32  032 -0.02991960  1.0288164318
#> 33  033  0.39629920  0.7783139307
#> 34  034 -0.58397583  0.0702545116
#> 35  035  1.29949045  0.5556062079
#> 36  036  1.43818779  1.8630581210
#> 37  037 -0.90947951  0.2549225079
#> 38  038 -0.18866102 -1.2718179597
#> 39  039  0.69822634 -0.2590574617
#> 40  040  0.80375182  0.2755621628
#> 41  041  1.91012994  1.1607094777
#> 42  042  2.07120146  1.7218456481
#> 43  043  0.27351347  0.2358145085
#> 44  044  0.21557616 -0.5009379498
#> 45  045  0.45694258  0.2255788342
#> 46  046  0.07189418 -0.9109294706
#> 47  047  0.75413816  0.5815285296
#> 48  048 -1.58794082 -2.6313513079
#> 49  049 -0.84379810 -0.1815528933
#> 50  050  2.81758158  1.7279582187
#> 51  051 -1.02296031  0.5455550057
#> 52  052  0.59998922  0.3784767936
#> 53  053  0.73634571 -0.8111604077
#> 54  054  1.12732982  0.9130484986
#> 55  055  0.83891381  0.8764081858
#> 56  056  0.95685032  0.4117615582
#> 57  057  0.64149407  0.3863612158
#> 58  058  0.30066470  0.7791292095
#> 59  059  2.20908826  2.4156854328
#> 60  060  0.81944623  0.7609882840
#> 61  061 -1.08876001  0.9281422969
#> 62  062 -0.23385053  0.1292720914
#> 63  063  0.01751318 -0.4703462714
#> 64  064 -0.31906324 -0.6325316448
#> 65  065  0.78810898  0.0318237515
#> 66  066  0.06948271 -1.4448749822
#> 67  067 -0.09292048  2.2835384962
#> 68  068 -0.36090102 -0.4995227581
#> 69  069  1.63883604 -0.3683380367
#> 70  070 -0.08626945 -0.8079701907
#> 71  071 -0.61591356 -1.4559097712
#> 72  072 -0.49528141 -0.6535364229
#> 73  073 -1.90507905 -1.9599359543
#> 74  074  0.01859816  1.6400021685
#> 75  075  0.39720771  1.2453882152
#> 76  076  0.07247160 -0.5591342259
#> 77  077 -0.05898392 -0.5596374748
#> 78  078  0.92161381  1.0773632950
#> 79  079  0.10501088  0.7679922381
#> 80  080  0.33236946 -0.1561102132
#> 81  081  0.81847455  0.1288881572
#> 82  082 -1.31977549  0.4015402695
#> 83  083 -0.66266269  0.7314084452
#> 84  084  0.84446659  1.8367842978
#> 85  085 -1.05601461 -1.4161784459
#> 86  086  1.29506013  0.5598081322
#> 87  087  0.43997303 -0.1063644211
#> 88  088 -0.47957655  1.3429879155
#> 89  089 -0.37161516  1.6027010051
#> 90  090  0.51092522  0.6078727123
#> 91  091  0.82914060  0.0139891589
#> 92  092  0.01755828 -1.4109828805
#> 93  093 -1.78958547 -0.6280928737
#> 94  094  1.98789861 -1.0059044330
#> 95  095 -0.96517720 -1.3501822804
#> 96  096 -1.02132727 -2.4101475178
#> 97  097 -1.09726491 -2.2443255492
#> 98  098 -0.27654886 -1.3745226539
#> 99  099 -0.55728590 -0.0005173414
#> 100 100 -1.84811854 -0.2210930007

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.94014633 -0.96700891
#> 2   002 -0.12833558  0.42390087
#> 3   003  0.16778314  0.24985287
#> 4   004 -0.65997557  0.74698459
#> 5   005  0.46296471  0.51919957
#> 6   006 -1.90623765 -1.02604030
#> 7   007  1.09518771  1.32612821
#> 8   008 -1.43553695  0.81096979
#> 9   009  0.74448391  0.85485294
#> 10  010 -0.30701617 -1.58591855
#> 11  011 -0.78245662 -0.32927147
#> 12  012  0.01749010  0.22047959
#> 13  013  0.54292186 -0.94401400
#> 14  014 -0.83315228  0.27652907
#> 15  015 -0.85053845 -0.86009842
#> 16  016  0.13891256 -0.69286269
#> 17  017  0.23605860 -1.50883060
#> 18  018  1.42597703  0.20887050
#> 19  019  2.72582424  1.58757602
#> 20  020 -0.45315916  0.67454193
#> 21  021  1.41883782 -0.01291782
#> 22  022  0.19885501  0.11436174
#> 23  023 -2.68203899 -0.36797692
#> 24  024 -0.18516861 -0.49757005
#> 25  025  1.66743264  1.52679052
#> 26  026 -0.06218915  1.42460659
#> 27  027 -1.42395751 -0.29157847
#> 28  028 -0.96182137 -2.81755964
#> 29  029  0.07594621 -0.81243651
#> 30  030  1.29295107 -1.05787959
#> 31  031 -0.32395870 -0.92093870
#> 32  032 -0.87169999  0.23873073
#> 33  033  0.77879108  1.20396538
#> 34  034 -1.79665725 -0.79490002
#> 35  035 -0.67174004 -1.17588033
#> 36  036  2.15708458  1.64617068
#> 37  037 -1.53687154 -1.70859232
#> 38  038  0.49780884 -0.73751480
#> 39  039 -0.39046785  1.22193820
#> 40  040 -1.40474250 -2.01736614
#> 41  041  1.25966483  1.44054114
#> 42  042  0.69497397  0.23800330
#> 43  043  0.07192279 -0.09966326
#> 44  044 -0.52102361  0.33845518
#> 45  045 -0.73458207 -0.26518890
#> 46  046  0.53211008  1.20251544
#> 47  047 -0.34562800  1.15260248
#> 48  048 -1.26467816  0.11853015
#> 49  049  0.24755617 -1.17462610
#> 50  050  1.59831666  1.96756745
#> 51  051 -0.25935203  0.34313346
#> 52  052  0.55304190 -0.09639827
#> 53  053  0.15695977  1.13091760
#> 54  054 -0.47212741 -0.85311245
#> 55  055  2.01921568  0.29982953
#> 56  056 -1.26172480  1.49543375
#> 57  057 -0.27907302 -0.24575961
#> 58  058 -0.45865805 -0.34845121
#> 59  059  0.10582600 -1.00570543
#> 60  060  1.82170414  1.01099145
#> 61  061 -1.78173245 -0.95476559
#> 62  062  1.17854686  0.97854608
#> 63  063 -0.56642663 -1.00616037
#> 64  064 -0.56244761 -0.22645691
#> 65  065 -1.52612227 -1.69175115
#> 66  066 -0.61596384  0.08859937
#> 67  067  1.39372690  0.02334848
#> 68  068  1.08689730 -0.24819075
#> 69  069  0.29196265  0.75424616
#> 70  070 -1.23321206 -0.68348003
#> 71  071 -1.09006280 -0.60893079
#> 72  072  0.42396464 -0.70831169
#> 73  073  0.19697726 -0.77394830
#> 74  074  0.96750053 -0.33886126
#> 75  075  0.43962094 -1.73283525
#> 76  076  0.35297867  1.52793218
#> 77  077 -0.38994858 -0.08752130
#> 78  078 -1.08671568 -1.28053555
#> 79  079  0.81288052  0.57773529
#> 80  080 -1.63246653  1.71865404
#> 81  081  1.20242615  1.36426617
#> 82  082  0.29536461 -0.19136717
#> 83  083 -1.06614128  0.29329257
#> 84  084  0.51461374  0.07364300
#> 85  085 -0.30078820  0.01091692
#> 86  086  1.11165532  0.19941812
#> 87  087  0.69705657  0.99634971
#> 88  088  2.32621466  2.35096130
#> 89  089  0.01254030  0.85839965
#> 90  090  0.17904314 -0.62682333
#> 91  091 -0.20611345 -0.92083465
#> 92  092  0.96424682  1.29845103
#> 93  093  0.59552868 -0.76089343
#> 94  094 -0.83778566 -0.16577707
#> 95  095 -0.17235082  0.15410057
#> 96  096 -1.36003620 -0.31756151
#> 97  097  0.94048578  0.34221603
#> 98  098 -0.62362175  0.84260052
#> 99  099  0.67214665  0.97858640
#> 100 100  0.39159827 -0.11800103