Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>          Y_1      Y_2
#> Y_1 1.000000 0.412131
#> Y_2 0.412131 1.000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.6884367  0.15930039
#> 2  2.0413216  1.49808229
#> 3  0.3986513 -0.07730195
#> 4  0.6430664 -0.18264564
#> 5  0.4740998  0.90411923

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.330064503 -1.91358243
#> 2   002 -2.568095537 -1.63584279
#> 3   003 -0.976491422  0.62060514
#> 4   004  0.512409445 -0.76577822
#> 5   005  0.189717406  0.36556211
#> 6   006 -0.120637460  1.01216634
#> 7   007  0.811467090 -0.62119215
#> 8   008  0.211462370  1.53110614
#> 9   009 -0.411598915 -2.13392135
#> 10  010 -1.767746557 -1.22973235
#> 11  011  0.848591647  0.80873142
#> 12  012 -2.178501353 -1.84081008
#> 13  013 -1.106181292  0.52893971
#> 14  014  1.229437829 -0.64952641
#> 15  015  0.557233678 -0.07172359
#> 16  016  1.806674725  1.08817731
#> 17  017  1.049571168 -0.65591024
#> 18  018 -0.158522809  0.38688411
#> 19  019  0.454370330 -1.51445798
#> 20  020  1.148270699 -0.16571825
#> 21  021  1.593676461  1.37524795
#> 22  022 -0.244347897 -1.18195072
#> 23  023 -0.343470848 -1.38418855
#> 24  024  0.796338228 -0.55667385
#> 25  025 -1.040704189 -0.10421737
#> 26  026  1.543495658  1.33124215
#> 27  027  0.197176896 -0.41447643
#> 28  028 -0.040385751 -0.17031300
#> 29  029  0.635502641  1.23252961
#> 30  030 -0.187384032 -0.51902949
#> 31  031 -1.340699277  0.16137977
#> 32  032 -0.293549566  0.77012475
#> 33  033  1.311760371  0.34394572
#> 34  034 -0.220453970  0.44596475
#> 35  035  0.116977253 -0.71721920
#> 36  036  0.183089484 -0.59181080
#> 37  037  0.626357424  1.38809449
#> 38  038  0.174422605  0.08108341
#> 39  039  1.369875750  2.23689302
#> 40  040  1.223448758 -0.11903659
#> 41  041 -0.075937851 -0.86245084
#> 42  042 -1.029563569  0.14355311
#> 43  043 -1.621152201 -0.21920129
#> 44  044 -1.394765599 -0.36527291
#> 45  045 -1.096677470  0.15684194
#> 46  046 -2.004845405 -0.56415908
#> 47  047  0.679402577  0.89647906
#> 48  048  0.664606798 -0.17259856
#> 49  049  0.312675668  0.38204901
#> 50  050  0.176443809  0.42804435
#> 51  051 -0.436305104 -1.51076194
#> 52  052  0.723361285  0.49631826
#> 53  053  0.470431410  0.17457808
#> 54  054 -0.534550087  0.13156855
#> 55  055  0.215518020  0.18744410
#> 56  056 -0.452229686  0.60084033
#> 57  057 -0.439089272 -1.15397732
#> 58  058  0.917617070  1.83716980
#> 59  059 -0.648367686 -0.17644233
#> 60  060 -0.537419832 -1.77330435
#> 61  061  0.850215996  0.84652008
#> 62  062  1.246690103  0.76767620
#> 63  063 -1.296997363 -1.59999820
#> 64  064 -0.438491773 -0.24166868
#> 65  065 -0.004336470 -0.00959160
#> 66  066 -0.208240242  0.04125640
#> 67  067 -0.009257335  0.30883632
#> 68  068  0.232511133  0.04497364
#> 69  069  0.778770297  0.40099472
#> 70  070  0.048096192  0.47635379
#> 71  071 -0.739107437  0.15278629
#> 72  072 -0.707700166 -0.19060980
#> 73  073 -1.894560131 -2.07256021
#> 74  074  1.311065582  0.50173510
#> 75  075 -1.930661343 -0.66577670
#> 76  076 -0.502667873 -1.06868834
#> 77  077  1.859246428  0.64466357
#> 78  078 -0.048541835 -0.46828867
#> 79  079 -1.716226773 -0.87107419
#> 80  080  0.315276943  0.46563483
#> 81  081 -0.869249373 -0.05200653
#> 82  082  1.379063329  1.08848577
#> 83  083  0.191279506  0.14455836
#> 84  084 -0.714467786 -0.59685310
#> 85  085 -1.252619013  0.04545659
#> 86  086  1.744584961  0.44342436
#> 87  087 -0.415440556 -0.79575283
#> 88  088 -2.055994637 -0.38264357
#> 89  089  0.430665146  0.66655901
#> 90  090  1.160689018  0.98830018
#> 91  091  0.126249619  0.30813236
#> 92  092 -0.260489750 -0.03070715
#> 93  093 -0.966418072  1.42302384
#> 94  094  0.339697931 -0.68741354
#> 95  095  0.457534120 -0.26983033
#> 96  096 -0.621143710  0.83966531
#> 97  097  0.439057284  0.08561226
#> 98  098  0.683237624  1.72619643
#> 99  099 -1.817825491 -0.53347875
#> 100 100  1.109595781  1.37972897

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.69450190  0.176136130
#> 2   002  0.51862071 -0.478382583
#> 3   003 -1.50716003 -0.197966417
#> 4   004 -0.19267907 -0.689576808
#> 5   005 -0.87087467 -0.564427091
#> 6   006  0.58953449 -0.183679484
#> 7   007  0.17471449 -0.163977892
#> 8   008  0.35626143 -1.326169745
#> 9   009  0.84543581  0.349890840
#> 10  010  0.53034993 -0.518635423
#> 11  011  1.29263030  0.424904286
#> 12  012 -0.49802925  0.165712301
#> 13  013  0.55315317  1.438834803
#> 14  014  0.19817082 -0.743033767
#> 15  015  0.67726184  0.091149083
#> 16  016  0.15796562  0.943664545
#> 17  017 -0.97075224  0.542893132
#> 18  018  0.92353182 -0.357069641
#> 19  019  1.25729509 -0.656470904
#> 20  020 -0.18765484 -0.433948942
#> 21  021  0.85609424 -0.229749185
#> 22  022 -2.25450177 -0.761863240
#> 23  023  0.27881563 -0.146404000
#> 24  024 -0.90725584 -1.176032592
#> 25  025 -0.97023165 -0.218879191
#> 26  026  1.71103602  1.838290851
#> 27  027  1.13049704  0.069965676
#> 28  028  0.32413995  0.576080037
#> 29  029  0.72072075  0.917958893
#> 30  030 -1.27045888 -1.115435113
#> 31  031  0.78233158  0.597932111
#> 32  032 -1.72038475  0.112313567
#> 33  033  0.84524807 -0.798093836
#> 34  034  1.44558194  0.014400490
#> 35  035 -0.05998533  1.139125757
#> 36  036  1.60929316  0.542556319
#> 37  037  0.86284735 -1.258441781
#> 38  038  0.70701682 -0.413713456
#> 39  039  2.68203784  1.673924604
#> 40  040  1.45209577  0.675156711
#> 41  041 -0.41333759 -0.721498323
#> 42  042  0.92349723  0.755147975
#> 43  043  0.08671112 -0.452847687
#> 44  044 -1.36107257 -0.515602322
#> 45  045  0.06808933  1.164309843
#> 46  046 -0.60441113  0.046151036
#> 47  047 -0.77535433 -0.870472320
#> 48  048 -1.39900260  0.341325757
#> 49  049 -0.42316633  0.461361721
#> 50  050  0.27869401 -0.266235287
#> 51  051  0.45412122  2.356842569
#> 52  052  0.72115657  1.146346262
#> 53  053  1.43338516  0.984128905
#> 54  054  1.26924461  2.302568105
#> 55  055 -0.75870593 -1.367396842
#> 56  056 -1.32193736 -1.098583345
#> 57  057 -0.83315174  1.352299109
#> 58  058  1.47070049  1.809068285
#> 59  059 -0.65151469  1.502319155
#> 60  060 -1.36019912 -1.611746532
#> 61  061  2.16737537  2.975785979
#> 62  062 -0.41034365  0.178514584
#> 63  063 -0.30075094 -0.445734902
#> 64  064  1.25983712 -0.824480398
#> 65  065 -0.70557091  1.602611517
#> 66  066 -0.47924464 -1.195893891
#> 67  067 -0.09362395 -0.347116667
#> 68  068  1.77178346  0.485935583
#> 69  069  0.68370098  1.626944316
#> 70  070 -0.54180049  0.012307115
#> 71  071  1.28395363  1.580673492
#> 72  072 -0.51755054 -0.122653311
#> 73  073 -0.08432956  2.332286889
#> 74  074  0.89082189  0.711531694
#> 75  075 -0.51246916 -0.932146810
#> 76  076 -0.53198412 -0.257796483
#> 77  077 -0.62245158 -0.294398497
#> 78  078 -0.64531105  0.198326287
#> 79  079 -0.76676740 -0.191577260
#> 80  080 -0.44324582 -1.268403617
#> 81  081  0.99440970  0.006289602
#> 82  082 -0.08540675  0.835302191
#> 83  083  0.39356131 -0.098103465
#> 84  084  0.29336093 -0.525291800
#> 85  085  0.79260881  1.212042308
#> 86  086  0.17941833  0.298594527
#> 87  087 -1.64680227 -0.701463487
#> 88  088 -1.17352919 -1.201531864
#> 89  089 -0.45890185 -1.563410092
#> 90  090  0.02812223  0.232336617
#> 91  091 -0.33079189  0.171112330
#> 92  092  0.83639999  3.138967126
#> 93  093 -0.43179354  1.437093902
#> 94  094 -0.67719282  0.282563891
#> 95  095  0.60529431  1.132855816
#> 96  096 -0.67681715  0.457533422
#> 97  097  0.95052078  0.458108123
#> 98  098  1.95971193  1.178865646
#> 99  099  1.06703153 -0.693113200
#> 100 100  1.41056396  1.072391448