Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4516432
#> Y_2 0.4516432 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.5643161  0.7438231
#> 2 -0.0496809 -1.9052382
#> 3 -0.4969077  0.6695819
#> 4  0.3214273  0.3782560
#> 5 -0.2307540 -0.3274282

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.164178363  0.821087381
#> 2   002 -0.177761431  1.704925034
#> 3   003 -0.936679959 -0.433297253
#> 4   004  0.856338462  0.943118025
#> 5   005  1.616867023  1.148847033
#> 6   006 -0.902561784  1.082054716
#> 7   007 -0.103771734 -0.381445015
#> 8   008 -0.656282315 -1.358032230
#> 9   009  0.163511702  1.158331946
#> 10  010 -1.367961021 -1.703562451
#> 11  011  0.240879555 -0.167582633
#> 12  012  0.958753784  0.572433993
#> 13  013  0.529256019  0.407225345
#> 14  014 -0.433411247 -0.379692383
#> 15  015 -0.544768002 -0.232969730
#> 16  016 -1.166110869 -1.960623596
#> 17  017  0.368112943  2.269382909
#> 18  018  0.674526791  1.382246111
#> 19  019 -1.389759828 -0.443792003
#> 20  020 -0.450609858 -0.491133913
#> 21  021  1.162842232 -0.268511878
#> 22  022 -1.331462614 -2.357689004
#> 23  023  0.615348727  0.959068541
#> 24  024 -0.014843838 -0.643992847
#> 25  025  0.768513330 -0.242090034
#> 26  026 -1.377807380  0.606600479
#> 27  027  0.103764235  0.736720763
#> 28  028  0.148852379 -2.197760523
#> 29  029 -0.409384932  0.314173462
#> 30  030 -0.394351994 -0.081830606
#> 31  031 -0.579937325 -0.979712388
#> 32  032 -0.464843081 -0.733463614
#> 33  033  0.233805347 -1.296029375
#> 34  034  0.731173508  0.190061986
#> 35  035  0.779748772 -0.439642400
#> 36  036 -0.944210214 -0.526882750
#> 37  037  0.036794989 -0.205114897
#> 38  038 -1.449843485 -0.039915365
#> 39  039 -1.342528660 -2.007595796
#> 40  040 -0.517019218 -0.195787559
#> 41  041  0.063894651 -0.820147143
#> 42  042 -1.825442522 -1.556551303
#> 43  043 -1.521097756  0.521956679
#> 44  044  0.665684220  0.827363008
#> 45  045 -1.015031285 -0.619284254
#> 46  046  0.607608666  1.826117276
#> 47  047 -0.448391116  1.754357461
#> 48  048  1.045387087  0.676982347
#> 49  049 -0.130692607 -0.041998180
#> 50  050  0.588939542 -0.961505018
#> 51  051  0.058711678  0.368585348
#> 52  052  1.041629569 -0.878557990
#> 53  053 -1.612988121 -0.073905076
#> 54  054 -0.671126906 -0.050728262
#> 55  055 -0.303877070  0.657002564
#> 56  056  0.211443817  2.205146698
#> 57  057 -1.458481655 -1.097912753
#> 58  058  0.119061241 -1.317118480
#> 59  059  0.197787256  0.825090225
#> 60  060  0.308065594 -1.955451968
#> 61  061  2.086050791 -0.168418694
#> 62  062  0.795226749  0.645092239
#> 63  063 -0.979288992 -0.266274537
#> 64  064 -0.866701755 -1.171934697
#> 65  065 -1.268955713  0.007200019
#> 66  066  1.346870245  0.546681398
#> 67  067  0.391306756  1.468762740
#> 68  068  0.581910757 -0.937492079
#> 69  069 -1.273621633 -0.761579094
#> 70  070 -0.693637626 -0.076108908
#> 71  071 -0.094799993 -0.855653872
#> 72  072 -0.930636808 -1.739785805
#> 73  073  2.553838729  1.331981243
#> 74  074  0.617554678  0.093103427
#> 75  075 -0.421929678  0.551425775
#> 76  076 -0.002577481  0.194352222
#> 77  077  0.224978741  0.330203626
#> 78  078  0.302751150 -0.541268995
#> 79  079 -0.599572809  0.540586043
#> 80  080 -0.615261204 -1.437188041
#> 81  081 -0.481813552  0.429391924
#> 82  082  0.472663503 -0.572397997
#> 83  083 -0.665221954 -1.150330104
#> 84  084  1.709281605  1.275532776
#> 85  085 -0.811958039  0.048086219
#> 86  086  0.240805193 -0.792337224
#> 87  087  0.904832012 -0.784499245
#> 88  088 -1.646616620 -2.204188110
#> 89  089  0.525091737 -0.573286027
#> 90  090  0.319310109 -0.753185763
#> 91  091 -0.326300755 -0.443794791
#> 92  092  2.171120287  2.908616693
#> 93  093 -2.396731455 -1.465900588
#> 94  094 -0.287706001  1.005194701
#> 95  095 -0.439875943 -0.177174854
#> 96  096 -1.247653689 -1.786612274
#> 97  097  0.405351964  0.335673121
#> 98  098  0.472556883 -0.481064667
#> 99  099  0.751648510  0.947672257
#> 100 100 -1.386670962  0.551604966

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -2.12022311 -0.863471522
#> 2   002 -0.02763701  0.686100354
#> 3   003 -1.14035099 -0.544668081
#> 4   004  1.24757853 -0.261470827
#> 5   005 -1.26600794  1.140840047
#> 6   006  0.10632273  0.172937513
#> 7   007  1.37799901 -0.206119588
#> 8   008 -0.43000007 -0.220706680
#> 9   009 -1.83329526 -0.561836119
#> 10  010 -0.29708234 -0.323686670
#> 11  011 -1.16071829 -0.592841851
#> 12  012  0.75763485 -0.788644135
#> 13  013  0.32158256  1.910316369
#> 14  014  0.15685081  0.547062017
#> 15  015 -0.71186734 -0.141262784
#> 16  016 -0.17216949 -0.358310178
#> 17  017  2.22514353  0.464632947
#> 18  018  0.24417070  0.888039771
#> 19  019 -2.18039464 -0.749737196
#> 20  020  0.11475862  0.403198945
#> 21  021 -1.18172352 -1.058864885
#> 22  022 -0.82252749  0.107073262
#> 23  023 -0.33180064  1.248843185
#> 24  024 -0.55845602  1.117993724
#> 25  025  1.13362171  0.042318897
#> 26  026  0.67438578  0.001608601
#> 27  027 -0.43583714 -1.172798651
#> 28  028  0.97804768 -1.000425958
#> 29  029  0.16201073 -0.273335614
#> 30  030 -0.18632935  0.264594321
#> 31  031  0.06047587 -0.629861643
#> 32  032  0.28663862  1.136088227
#> 33  033 -1.32307220 -1.412002331
#> 34  034 -0.30561699  0.104911158
#> 35  035  0.14999849 -0.921094870
#> 36  036  0.33495684  0.605724740
#> 37  037  0.26732456  0.533140092
#> 38  038  0.25319146  1.684541862
#> 39  039  0.43784400  0.429770441
#> 40  040  0.73745585  0.167207335
#> 41  041  0.07858582  0.773423957
#> 42  042 -0.76445976  0.574962975
#> 43  043 -0.04865984  0.511074115
#> 44  044  1.82788813  1.443596369
#> 45  045 -2.21187312  0.025097298
#> 46  046  2.08872541  0.893313712
#> 47  047 -0.57204811 -1.428648600
#> 48  048  0.24850595 -1.049972411
#> 49  049  1.40295149  0.123738500
#> 50  050 -1.28266463 -0.227983750
#> 51  051  1.06404680 -0.310427819
#> 52  052 -1.28829935 -0.526650484
#> 53  053 -1.38900454 -0.953414767
#> 54  054  2.16647111  1.787236105
#> 55  055 -0.34470886 -0.064504634
#> 56  056  0.71545750  0.467130298
#> 57  057  0.16831658  0.889171347
#> 58  058 -0.77380677 -0.796284032
#> 59  059 -0.42239162 -0.218104221
#> 60  060 -0.35269788  0.200056626
#> 61  061  0.45857277  1.010382555
#> 62  062 -0.60603523 -1.454319846
#> 63  063 -0.53808218 -1.697579071
#> 64  064  0.19055620  0.195719566
#> 65  065  0.82091135 -1.736339898
#> 66  066 -0.13400780 -2.049791330
#> 67  067 -0.37795243 -0.187653753
#> 68  068  0.63862941  1.157047740
#> 69  069  0.08082341 -1.276779649
#> 70  070 -2.87383021 -0.648496617
#> 71  071  0.71653918  0.756559169
#> 72  072  0.42626857  0.326356433
#> 73  073  0.30880844 -0.021526717
#> 74  074  0.43690949  0.452752798
#> 75  075  0.26553158  2.240789964
#> 76  076 -0.51262549 -0.885174590
#> 77  077 -0.19071553  0.908568761
#> 78  078 -0.02401451 -0.486613014
#> 79  079  0.81436425  0.002467661
#> 80  080 -1.32412882 -0.665798180
#> 81  081  0.47878404  0.116606886
#> 82  082  0.60008518  0.725869829
#> 83  083  2.24075781 -0.464866639
#> 84  084  0.41325061  1.121093237
#> 85  085 -0.58371902 -0.587823363
#> 86  086  1.96816422  1.823590918
#> 87  087  1.65406384  1.513214590
#> 88  088  0.68605755  0.051969792
#> 89  089 -0.53063347  0.134656631
#> 90  090  0.70327299 -0.764658976
#> 91  091 -0.86555153 -0.458057096
#> 92  092 -0.51158973 -1.193862968
#> 93  093  0.77512659  0.441945356
#> 94  094 -0.33227311  0.244867746
#> 95  095 -0.21486750  0.185434044
#> 96  096 -1.75637212 -2.301338085
#> 97  097 -0.64170855  0.441907060
#> 98  098 -1.04539908  0.767981754
#> 99  099  0.12916136 -0.286460733
#> 100 100  0.28677697  0.101662919