Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4938615
#> Y_2 0.4938615 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.8134677 -1.06209186
#> 2 -1.2227203 -1.45646849
#> 3 -0.1776317  0.67808452
#> 4 -1.9717452 -0.03044479
#> 5  0.3510273  0.31310295

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.56167319 -0.816555662
#> 2   002 -0.73581053  0.454325038
#> 3   003 -0.99588095 -0.237300417
#> 4   004 -0.61348254 -0.176175482
#> 5   005 -0.51314348 -0.149018466
#> 6   006 -1.25751520 -2.000705346
#> 7   007  1.16679268 -0.864340815
#> 8   008  0.42287542 -0.276451081
#> 9   009 -0.80384106  0.878800734
#> 10  010  0.12674641 -0.973444950
#> 11  011 -0.49751107 -1.930441548
#> 12  012 -0.71420787 -0.032972671
#> 13  013 -0.30813744  0.544497797
#> 14  014 -0.18514716  1.010684362
#> 15  015 -1.68660542 -0.757505900
#> 16  016 -1.07645642 -1.104059349
#> 17  017  0.47257677 -0.323614604
#> 18  018  1.60769305  1.570390163
#> 19  019 -0.18078859 -1.797844951
#> 20  020  0.91240025 -0.139717563
#> 21  021 -2.07628681 -0.311865288
#> 22  022 -0.40283841 -0.513717795
#> 23  023  0.49011529  1.149834104
#> 24  024 -0.04285942  1.748672789
#> 25  025 -0.82803526 -0.113265660
#> 26  026  0.20338675  0.248813288
#> 27  027 -0.46041757  0.206026045
#> 28  028 -0.69615805  0.818136208
#> 29  029 -1.39754661 -0.872657070
#> 30  030 -0.71504484 -1.297490546
#> 31  031 -1.43518838 -0.553136536
#> 32  032  2.36623087  1.572481997
#> 33  033 -0.73819836 -0.100828145
#> 34  034 -0.63193524  0.114696728
#> 35  035  1.16523305  1.036951856
#> 36  036 -1.45277794 -1.878824723
#> 37  037  0.04384747  1.059869490
#> 38  038  0.50769511  0.124790101
#> 39  039  0.80142660 -0.542219564
#> 40  040  0.30517043  1.235427316
#> 41  041 -0.21510751  0.816027408
#> 42  042 -1.28776317 -0.903429813
#> 43  043  0.96979033  0.619871647
#> 44  044 -0.47945098  0.633713619
#> 45  045 -1.64857398 -1.581692245
#> 46  046  0.38081164  0.479069448
#> 47  047 -0.41536580 -1.431431625
#> 48  048  0.04609157  0.832234417
#> 49  049 -1.73978536 -2.787932056
#> 50  050 -1.52930723 -1.579312448
#> 51  051 -1.52017606 -1.162782803
#> 52  052 -1.56669292 -2.040984241
#> 53  053 -1.92728970 -2.247002076
#> 54  054  0.37231062  0.040347623
#> 55  055 -0.52350103 -1.122837374
#> 56  056 -1.40828156  0.664418267
#> 57  057  0.80189538 -0.370965275
#> 58  058 -0.71591301 -0.453629759
#> 59  059  0.86067449  0.682584863
#> 60  060  0.90251167  0.546991221
#> 61  061  0.58693692  0.278817938
#> 62  062 -0.29843851 -0.552488251
#> 63  063  2.06421733  1.011842272
#> 64  064  0.27678482 -0.597274615
#> 65  065 -0.33658035  1.246898665
#> 66  066  0.12004126 -0.402743271
#> 67  067  0.78269329  1.026242464
#> 68  068  0.20766167 -1.081372756
#> 69  069  1.49001150  0.134587537
#> 70  070 -0.61468609  1.084025507
#> 71  071  0.20363456 -0.108283279
#> 72  072  1.67950351  1.046229384
#> 73  073  0.38634631 -0.536289205
#> 74  074  0.55263208 -1.582511140
#> 75  075 -0.08969108 -1.666869045
#> 76  076 -0.16775778 -1.134982473
#> 77  077  0.29753080  0.401386940
#> 78  078 -0.77211640 -0.862254305
#> 79  079 -1.53552589  0.242860760
#> 80  080 -0.02418541 -1.114822290
#> 81  081 -1.33948106 -1.095532480
#> 82  082  0.10121751  1.202300604
#> 83  083 -2.68089359 -0.164879309
#> 84  084  0.79591741  1.175060609
#> 85  085  0.68159745  0.111555567
#> 86  086 -1.38910196 -0.071297326
#> 87  087  0.70823467  1.088954630
#> 88  088  1.35213468 -0.219134126
#> 89  089  0.32413672  0.588707779
#> 90  090 -0.65599111  0.338790410
#> 91  091 -0.97581363  0.117768903
#> 92  092 -0.20350702 -0.691198396
#> 93  093  0.54847686  0.070310670
#> 94  094 -0.39941814 -0.473584216
#> 95  095  1.67280767  1.147372862
#> 96  096 -1.84995457  0.888570861
#> 97  097 -0.63689357 -0.139562648
#> 98  098 -0.23286779  0.298284787
#> 99  099  0.56707702 -0.008528869
#> 100 100  1.37880016  0.661377735

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.97601539 -0.13038094
#> 2   002 -1.42094351 -0.42086040
#> 3   003 -0.16752667 -0.46739820
#> 4   004 -0.07920974  0.35900650
#> 5   005  3.25443574  0.63101961
#> 6   006 -0.77978301  0.01855271
#> 7   007 -1.54567473 -0.39901731
#> 8   008 -0.31195110 -0.71975138
#> 9   009 -1.13257560  0.41818754
#> 10  010  1.72119829  1.27170189
#> 11  011 -0.95912234 -1.81380840
#> 12  012  0.96861840  0.85580359
#> 13  013  0.30300403  1.01281630
#> 14  014 -1.48120028 -0.49668917
#> 15  015 -0.29025667  0.02834660
#> 16  016  0.39460335  1.45620812
#> 17  017 -0.29226076 -0.91542814
#> 18  018 -1.02402004 -0.23609606
#> 19  019  0.80381474  1.04314285
#> 20  020 -0.88032010  0.73265035
#> 21  021 -0.73529347 -0.33934838
#> 22  022  1.66301358 -0.19678642
#> 23  023 -0.12842009  0.37305772
#> 24  024 -0.26956510 -0.96978151
#> 25  025  0.42748010  0.06888544
#> 26  026 -0.29890989  0.21905410
#> 27  027  0.44165813  0.60049730
#> 28  028  0.86179539  1.14249572
#> 29  029  0.03646487 -0.33279006
#> 30  030 -1.76362047 -0.92435575
#> 31  031  0.33150545  0.71083029
#> 32  032  1.55340829 -0.05356786
#> 33  033  2.46125194  1.33278225
#> 34  034  0.03183361  0.02925200
#> 35  035  1.34429947  1.15673532
#> 36  036 -0.14407494 -0.45977660
#> 37  037  0.72766042  0.07660988
#> 38  038 -0.16497669 -1.84936002
#> 39  039  0.95454178  0.15281193
#> 40  040 -0.23719924  0.39133135
#> 41  041 -0.30731319 -1.28757310
#> 42  042  0.42677488 -1.40068829
#> 43  043 -0.68608613 -0.95773747
#> 44  044 -0.11819315 -0.73209125
#> 45  045  1.67795971  0.69291630
#> 46  046  0.92373798 -0.02919056
#> 47  047 -0.70793488  0.25245558
#> 48  048  0.32727949  1.50584709
#> 49  049 -0.06335742 -0.58868805
#> 50  050  1.54058712  1.98500730
#> 51  051  0.86223591 -0.92524151
#> 52  052  0.89746893 -0.85905675
#> 53  053 -0.13283097  0.08175257
#> 54  054 -1.57601527 -1.82843696
#> 55  055 -0.02815077 -0.16476219
#> 56  056 -1.14194221 -0.36208116
#> 57  057 -0.13093491  0.22499926
#> 58  058  1.70393370  1.12937959
#> 59  059 -2.02021582 -1.47206197
#> 60  060  0.78128707  0.74452661
#> 61  061 -0.91028088 -0.46722207
#> 62  062  1.17076484  0.77333465
#> 63  063 -0.78926383 -0.33848848
#> 64  064  1.13566631  1.61559365
#> 65  065 -1.68019307  0.56822939
#> 66  066  0.41390028  0.04249888
#> 67  067 -1.18198041  0.82090749
#> 68  068 -0.38155812  0.16355322
#> 69  069  0.48789727 -1.58318979
#> 70  070  1.04715286  0.29345908
#> 71  071  0.20748190  0.03998170
#> 72  072  1.67421734  0.60091586
#> 73  073  0.18017501  0.67111803
#> 74  074  0.23603497 -0.32892637
#> 75  075  0.84313601 -0.48298312
#> 76  076 -1.21322859 -0.04814289
#> 77  077 -1.54059638 -0.74771554
#> 78  078 -0.45851487  0.26200214
#> 79  079 -0.58730564 -0.20095335
#> 80  080  1.32040467 -0.76393114
#> 81  081  0.62854612  0.26186271
#> 82  082  0.13581762 -0.85390838
#> 83  083 -0.01995012  0.69312832
#> 84  084  1.03331942 -0.81196726
#> 85  085  0.99620818  1.04874061
#> 86  086 -0.43597571 -0.71031061
#> 87  087 -2.30051747 -1.17278587
#> 88  088  0.25299190 -0.35939503
#> 89  089 -2.16097158 -0.29129002
#> 90  090 -1.71289100 -1.77380307
#> 91  091  0.19811667  0.52111224
#> 92  092  0.50640863  0.06669452
#> 93  093  0.06866571 -0.30654035
#> 94  094 -0.37567444 -0.48578828
#> 95  095  1.20215261 -0.34386133
#> 96  096  0.38888690 -0.79650283
#> 97  097  0.03926810  1.55444666
#> 98  098  0.56943209  0.46034515
#> 99  099  0.69982518  0.89637058
#> 100 100  0.65217125  0.18130640