Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5207038
#> Y_2 0.5207038 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  1.3461998 -1.2888192
#> 2 -0.2651211  0.8534963
#> 3  0.3949781 -0.2580690
#> 4  0.4582114 -0.1848184
#> 5  1.5664786  1.6208100

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.69871872  0.56470725
#> 2   002  0.03525245 -1.06930854
#> 3   003 -1.17384327 -0.16260439
#> 4   004  0.64247381  0.83107880
#> 5   005 -0.92110080 -0.23993246
#> 6   006 -1.36662027 -1.87401705
#> 7   007  0.50588903  0.28863284
#> 8   008 -0.55238122 -0.22397427
#> 9   009  0.34041205 -0.12144803
#> 10  010 -1.42998568 -0.87889248
#> 11  011  0.21416065 -1.98400497
#> 12  012  0.26747830 -0.23611336
#> 13  013 -0.61206426  0.78634537
#> 14  014  0.62950243 -0.44553760
#> 15  015 -1.27455777  0.47122144
#> 16  016  0.07335636 -0.97095125
#> 17  017 -0.26836168 -1.28843287
#> 18  018  0.95971949  0.94626506
#> 19  019 -0.28318613  0.49654048
#> 20  020 -0.27001907 -0.30457183
#> 21  021  0.53964629 -0.23428018
#> 22  022 -0.10198725 -0.02413691
#> 23  023  0.17531229 -0.92648585
#> 24  024  0.72810838  1.16011104
#> 25  025 -0.25801926 -1.07271675
#> 26  026  1.96253716  2.59077981
#> 27  027  0.87507668  1.81213069
#> 28  028  1.07623288  0.97608477
#> 29  029  0.17329276 -0.14530484
#> 30  030  1.80768186  2.23889633
#> 31  031  0.03190461 -0.44489088
#> 32  032 -0.56123011 -1.52102438
#> 33  033  0.71492761  1.55618259
#> 34  034 -1.19774404 -1.26076605
#> 35  035  0.28667830  0.19153613
#> 36  036  0.11172086  1.30716298
#> 37  037 -0.92618416 -0.81718210
#> 38  038 -0.65249453 -0.33527077
#> 39  039  0.91763640  0.65701230
#> 40  040  0.31130629  0.55681107
#> 41  041  0.75628844  1.06861852
#> 42  042  0.42867557  0.12202275
#> 43  043 -1.89728830  0.77501240
#> 44  044 -0.33381343 -0.02721842
#> 45  045 -0.85586513 -0.76123885
#> 46  046  1.13945771  1.59266617
#> 47  047 -0.89363416 -0.01804095
#> 48  048 -0.34619665 -1.09384376
#> 49  049 -0.54913497 -1.10664842
#> 50  050 -1.35382124 -1.82924941
#> 51  051 -0.97143861 -0.95872544
#> 52  052  0.34838893  0.14355141
#> 53  053  0.13176318  1.59851961
#> 54  054  0.33762701 -0.73146599
#> 55  055  0.38023142  0.19280465
#> 56  056 -1.22542751 -1.07511039
#> 57  057 -0.15683829  1.08514646
#> 58  058  1.27396830 -0.25314533
#> 59  059 -0.11260963  0.09510062
#> 60  060 -0.89764269 -1.54662267
#> 61  061 -1.53838622 -0.07361172
#> 62  062 -0.07630251 -0.31425548
#> 63  063  0.91940955  0.24388998
#> 64  064 -0.04928485 -0.29702909
#> 65  065  0.10276834 -0.66647863
#> 66  066  2.52660634  1.50207673
#> 67  067  0.51850815 -0.25160899
#> 68  068 -0.02551698 -1.37926907
#> 69  069 -1.37535069 -0.44440000
#> 70  070 -0.50161715 -0.96211979
#> 71  071 -1.25420785 -1.11890370
#> 72  072 -1.31385509 -0.23132705
#> 73  073 -0.02780433  0.41342480
#> 74  074  1.62298080  1.42032386
#> 75  075  0.36192773 -1.04800051
#> 76  076  1.32766159 -0.01676795
#> 77  077 -1.26412921  0.31711838
#> 78  078  0.88559371  1.37606808
#> 79  079 -0.74352136 -0.41576065
#> 80  080 -2.26261814 -0.88388448
#> 81  081 -1.37759348 -0.48681924
#> 82  082  1.87399985  0.68075872
#> 83  083 -0.15210656  0.22566660
#> 84  084  1.47760942  0.96891840
#> 85  085  1.71347916  1.45302989
#> 86  086 -0.18129553 -0.04151636
#> 87  087  0.18055950  1.82957212
#> 88  088 -0.63587431 -1.71825563
#> 89  089  0.10455463 -0.27377307
#> 90  090 -0.24600303 -0.27683441
#> 91  091  0.75893430  0.66403041
#> 92  092 -0.18101762  0.08498253
#> 93  093  0.97809558 -0.05408722
#> 94  094  0.55305112  1.32083170
#> 95  095  0.55202704 -0.15689984
#> 96  096  0.41644629  0.79166505
#> 97  097  2.55013907  1.30483086
#> 98  098 -1.05304202 -1.00268214
#> 99  099  1.95799330  0.90426105
#> 100 100 -0.81101596 -0.55825802

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  0.024921810  0.104627712
#> 2   002 -1.334699584 -1.020079111
#> 3   003 -1.050908263 -0.207575958
#> 4   004  1.613907274  0.951455050
#> 5   005  0.319216153 -1.648412036
#> 6   006  0.430464885  0.140720018
#> 7   007 -0.285467819 -0.369869719
#> 8   008  2.128454396  1.211608917
#> 9   009 -0.267781568 -0.479565465
#> 10  010 -0.632644955 -1.176955305
#> 11  011  1.829883106  0.810352692
#> 12  012  0.048886739 -1.559575102
#> 13  013 -1.339515480 -1.987563429
#> 14  014 -1.103639623 -1.464284378
#> 15  015 -0.099065415 -0.982587506
#> 16  016  0.248774161 -1.259504073
#> 17  017 -0.464136646  0.133701509
#> 18  018  0.332166716  0.409939054
#> 19  019  0.626898864  0.913732121
#> 20  020  0.440403942 -0.593253500
#> 21  021 -0.309730703  0.492122828
#> 22  022  0.129650700 -0.519301855
#> 23  023 -3.540560048 -1.530597413
#> 24  024 -0.767964742 -1.571345769
#> 25  025 -0.793606651 -1.156646636
#> 26  026 -1.392305647 -2.084903573
#> 27  027  2.087087757  0.619529144
#> 28  028  1.638348742  0.763640500
#> 29  029  0.008721135  1.445198142
#> 30  030 -1.381515849 -0.076399768
#> 31  031  0.591146610 -0.308268195
#> 32  032  0.016101373 -1.123730654
#> 33  033 -0.089601784 -0.537667252
#> 34  034  0.217500418 -2.044653842
#> 35  035  1.233051082  0.984065905
#> 36  036  1.048220383  0.505005968
#> 37  037 -0.076801694  0.297423508
#> 38  038 -0.441064633 -1.859872161
#> 39  039  0.686839427  1.327156170
#> 40  040  1.245953296 -0.087293876
#> 41  041 -0.236732149 -0.996774122
#> 42  042  0.366360597  0.680793753
#> 43  043 -0.068005198  0.380742114
#> 44  044 -0.470031916  0.092606392
#> 45  045  0.262491866 -0.671015760
#> 46  046  0.213322767 -0.027935368
#> 47  047 -1.174925776  0.152018889
#> 48  048 -0.098605846 -0.531760418
#> 49  049  0.774893078 -1.701519358
#> 50  050 -0.636041988  0.209339220
#> 51  051  0.516009543  0.431259335
#> 52  052  1.663330497  1.316978608
#> 53  053 -0.136631081 -0.004409169
#> 54  054  0.017328966  1.095199531
#> 55  055  1.100631863  0.071395068
#> 56  056  0.697258691  0.835075007
#> 57  057  0.021826177 -0.898098964
#> 58  058 -1.396315119 -0.901753656
#> 59  059 -1.765039195  0.611012178
#> 60  060 -0.348201979 -0.002639049
#> 61  061  0.345742585 -0.720766182
#> 62  062  0.530187693 -1.007594601
#> 63  063 -1.841018621 -2.053542863
#> 64  064 -2.306083555 -1.537258612
#> 65  065  0.038904893  0.229595438
#> 66  066 -2.637917986 -1.779357139
#> 67  067  0.066547364 -0.303834360
#> 68  068 -0.923371482 -0.310090964
#> 69  069 -2.395360949 -1.151014057
#> 70  070  0.313572043 -0.692299555
#> 71  071  1.139573618  0.822310055
#> 72  072  1.098829862  0.397446498
#> 73  073 -1.574112553 -1.716786239
#> 74  074 -0.016882325 -1.249490396
#> 75  075  1.671202340  0.529325651
#> 76  076  1.568793028 -0.339842409
#> 77  077  0.464483828  0.100160806
#> 78  078 -0.476856534 -0.776615554
#> 79  079  0.832197920  0.825261310
#> 80  080 -0.079413431 -0.174395808
#> 81  081  0.819982203 -0.019087465
#> 82  082 -1.522973805 -1.825413014
#> 83  083 -0.117803787  1.435514426
#> 84  084 -0.286938054 -0.313979533
#> 85  085 -0.555548966 -0.020180449
#> 86  086 -1.438591846 -0.253738090
#> 87  087 -2.227761303 -0.344903837
#> 88  088 -0.133509714  0.046611796
#> 89  089 -0.756232029 -0.195670316
#> 90  090  0.366888937 -0.906817344
#> 91  091  0.276426305  0.587934639
#> 92  092  0.014623527 -1.059273076
#> 93  093 -1.677536258 -1.717434767
#> 94  094 -0.712140336 -0.234072775
#> 95  095 -0.634311637 -0.892145500
#> 96  096  1.126376226  0.195428848
#> 97  097  0.217519208  0.785627557
#> 98  098 -1.031465350 -1.213960760
#> 99  099 -1.663555924  1.227725541
#> 100 100 -0.727577044 -0.837583752