Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4909245
#> Y_2 0.4909245 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.6098334 -0.3520480
#> 2  0.9652662 -0.3736256
#> 3 -1.0644044 -0.6477709
#> 4 -1.7140410 -0.8588907
#> 5  0.4909324 -0.2509072

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.41602430 -0.38647932
#> 2   002 -0.26774848 -0.50406833
#> 3   003  1.20723933  0.55803178
#> 4   004  0.56183492  1.30601720
#> 5   005  2.33107041  0.65166351
#> 6   006 -0.76099819 -0.42975614
#> 7   007  0.94634394 -1.59950711
#> 8   008  0.54889008  0.32960026
#> 9   009 -0.51021358 -0.14185502
#> 10  010  0.15065884  0.61041526
#> 11  011 -0.24488906 -0.36085137
#> 12  012 -0.21823899  0.16192423
#> 13  013 -0.92985617  0.24011686
#> 14  014  1.72442263  0.07398711
#> 15  015 -0.56588131 -0.26133361
#> 16  016  0.31990381  0.34691794
#> 17  017 -0.77392602 -0.92281945
#> 18  018  2.14554649  3.13847505
#> 19  019  0.25187921 -0.45249743
#> 20  020 -1.77915199 -0.12270753
#> 21  021 -0.21520508 -0.67184566
#> 22  022  1.29496150  0.77234888
#> 23  023 -0.29862060 -0.39613118
#> 24  024 -1.12111865 -0.83660452
#> 25  025  0.15472739  0.16317413
#> 26  026 -0.86492906 -0.93922248
#> 27  027  0.02019118  0.86385001
#> 28  028  0.98302884  1.16431471
#> 29  029  0.06752836 -0.87511002
#> 30  030  0.07401516 -1.58101705
#> 31  031 -0.03527523 -0.08295212
#> 32  032  1.37716629  0.44906178
#> 33  033 -0.42638436 -0.61690165
#> 34  034  0.33041493  0.46305676
#> 35  035  0.36974098  1.02189315
#> 36  036 -0.74725049  0.38197296
#> 37  037  0.30836432  1.10183208
#> 38  038  0.29639542  2.05908055
#> 39  039  1.26541853 -0.66932850
#> 40  040  0.92344350 -0.07735261
#> 41  041  0.07300801 -0.43660348
#> 42  042  0.91498757  0.08556009
#> 43  043 -0.96840344 -1.59726184
#> 44  044 -0.47243732  2.02437326
#> 45  045  0.93805939  0.11727901
#> 46  046 -1.10672271 -0.10843897
#> 47  047 -0.44855931  1.12065722
#> 48  048  0.60506775  1.25102472
#> 49  049  1.23309720  0.45369852
#> 50  050  0.04152603  0.91362286
#> 51  051 -0.06474570 -1.41803553
#> 52  052 -0.75370858 -0.01874247
#> 53  053  0.99096573  0.84199216
#> 54  054  0.42763298 -0.77984115
#> 55  055 -1.17564646  0.51057863
#> 56  056  0.03128454 -0.57223105
#> 57  057 -0.99676480 -0.45441870
#> 58  058 -0.61905270 -0.97916335
#> 59  059 -0.95820510 -0.20615580
#> 60  060 -1.87466917 -0.74396045
#> 61  061  1.13755410  0.03257396
#> 62  062 -1.65157271 -1.24630182
#> 63  063 -0.13426763 -1.22985386
#> 64  064 -1.08270481 -0.72686891
#> 65  065 -0.82133542 -0.24789916
#> 66  066 -0.45267488  0.21065365
#> 67  067 -0.47070004 -0.10914587
#> 68  068  1.37499031  0.60519853
#> 69  069 -1.86338726 -0.71986941
#> 70  070  0.74583145  0.12464796
#> 71  071 -1.02599069  0.11764154
#> 72  072 -0.53133668  0.59849999
#> 73  073 -1.51562542 -0.28905589
#> 74  074  1.71525621  1.70463779
#> 75  075 -0.14384402  1.19573659
#> 76  076 -1.21219218 -1.45861347
#> 77  077  0.93701772  0.18570140
#> 78  078 -0.55421751  0.29381144
#> 79  079 -1.57607164 -1.26032834
#> 80  080  1.98021240  0.12222781
#> 81  081  1.14124888  0.91458371
#> 82  082 -0.55822073 -1.00451243
#> 83  083 -1.77598062 -0.75381446
#> 84  084  1.86483416  0.28696128
#> 85  085  0.40112318  1.27506405
#> 86  086 -0.45422357 -0.17774407
#> 87  087 -0.30219844  0.09161526
#> 88  088  1.64152569  0.69155752
#> 89  089  0.29855180 -0.23761221
#> 90  090 -0.39660962 -0.69508465
#> 91  091  0.31291138  0.69874500
#> 92  092  2.05360178  0.24873367
#> 93  093  1.83416669  2.14979900
#> 94  094 -0.09892943  1.24308171
#> 95  095  0.98457462  1.23422363
#> 96  096  0.28375677  0.29218313
#> 97  097 -1.75739592 -1.06004621
#> 98  098  0.28584562  0.32511827
#> 99  099  2.05806049  0.68171534
#> 100 100  0.16316854  1.01018860

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.11162508 -1.645322180
#> 2   002 -0.37113222  0.281076843
#> 3   003 -1.32222189 -0.772330229
#> 4   004  1.37820464  1.760665794
#> 5   005  1.38985317 -0.374075866
#> 6   006  0.33366474  0.596022119
#> 7   007  1.43323102  0.680690046
#> 8   008 -2.12500530 -0.015050058
#> 9   009 -0.97320551 -0.782022405
#> 10  010  1.00188150  0.507975385
#> 11  011  1.53455318  2.516294237
#> 12  012 -0.14586358 -1.296198397
#> 13  013  0.26643522  1.010834689
#> 14  014  0.02172998  0.532076780
#> 15  015 -1.16658778  0.163349717
#> 16  016 -0.18872711  2.099833881
#> 17  017 -1.42060609 -1.499013142
#> 18  018  0.45718580  0.147680200
#> 19  019  1.38917107  0.166952279
#> 20  020 -0.58170722 -0.310453707
#> 21  021  0.87751692  0.658028467
#> 22  022  0.52315533 -1.273808988
#> 23  023  0.53277702 -0.654554138
#> 24  024  0.58681320  0.124305624
#> 25  025  1.72921558  0.965539440
#> 26  026  0.68637712  1.531130379
#> 27  027 -0.37990513  0.602719296
#> 28  028 -0.14351554  0.349535074
#> 29  029 -1.30050721 -0.712646124
#> 30  030  0.21032316  0.513397367
#> 31  031 -0.49967288 -0.195219848
#> 32  032 -0.86000062 -0.115348888
#> 33  033  0.19730473  0.140664534
#> 34  034  0.86052436  1.545788159
#> 35  035  1.60273932  0.401946878
#> 36  036 -0.98094827 -1.650521765
#> 37  037 -0.65828554 -0.407047002
#> 38  038  1.10247654  2.799653629
#> 39  039  0.13309111 -0.672033048
#> 40  040 -0.15439922  1.626275446
#> 41  041 -0.19279991 -1.153808562
#> 42  042 -0.36787820 -0.579487541
#> 43  043  0.13971485 -1.152382261
#> 44  044  1.39843724  0.538179988
#> 45  045 -0.65863165  0.938796103
#> 46  046 -0.42157563  0.591665595
#> 47  047 -0.03574498 -0.434366175
#> 48  048  0.16182344  1.317355460
#> 49  049 -0.97823795 -1.355891409
#> 50  050  0.15896940  0.020426835
#> 51  051  0.55808034  0.847123015
#> 52  052  0.32295280 -0.477079992
#> 53  053 -0.74799408 -0.626950279
#> 54  054 -1.54359464 -0.235564623
#> 55  055 -1.78908425 -1.830334947
#> 56  056  1.54772554  0.439968275
#> 57  057  1.55146007  1.563821562
#> 58  058  0.80591085  1.755713731
#> 59  059 -2.14603518 -1.015771809
#> 60  060 -0.68912759  1.091489243
#> 61  061 -0.36993575 -0.189318050
#> 62  062  0.87890535 -0.043085750
#> 63  063  0.35061000  0.152889938
#> 64  064  0.55373056  0.952804743
#> 65  065 -2.78542833 -0.401672181
#> 66  066  1.97918772  1.845534647
#> 67  067 -0.39364138 -0.002906471
#> 68  068  0.58494397 -0.029260524
#> 69  069 -0.70888552  0.394067118
#> 70  070  0.53298624 -0.717268685
#> 71  071 -0.60370296 -2.053134246
#> 72  072  0.98609539  0.457190498
#> 73  073 -0.34500257  0.175925361
#> 74  074 -1.00863456 -1.375969422
#> 75  075 -1.48312166 -0.828562291
#> 76  076  0.16873112  0.715681549
#> 77  077  0.80428897 -0.004298807
#> 78  078  0.48247739 -0.350954866
#> 79  079  0.75514304 -0.192636922
#> 80  080 -1.13711231 -0.558598643
#> 81  081  0.65514835 -0.291628729
#> 82  082 -1.25484974 -1.730379554
#> 83  083 -0.56763350  0.829624884
#> 84  084  0.40446719  0.468347258
#> 85  085  0.05060925  0.954088388
#> 86  086 -0.23624921 -1.036750863
#> 87  087  0.53477599 -0.588339876
#> 88  088  2.30376236  0.602740952
#> 89  089  0.75204578 -0.186218016
#> 90  090  0.53117215  1.079935474
#> 91  091  0.36755916 -0.772063299
#> 92  092  0.34159236 -0.936446939
#> 93  093  0.67603246 -0.448961637
#> 94  094  0.38746668  0.153922051
#> 95  095  1.22989029  2.336883764
#> 96  096 -0.62503520  0.660036910
#> 97  097 -1.15398823 -0.492236196
#> 98  098  0.50384349 -0.686705903
#> 99  099 -0.36048549  0.079691324
#> 100 100 -1.73575153 -0.510354931