Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5058676
#> Y_2 0.5058676 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.8687172 -0.3905905
#> 2 -0.1748995 -0.4316932
#> 3 -0.2014302  0.2549705
#> 4 -0.4390362 -1.1930508
#> 5 -0.4550297  0.1960311

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.45914761  0.24508360
#> 2   002  0.28890811  0.34142405
#> 3   003  0.63330575  0.67824905
#> 4   004  0.04933472  0.93422180
#> 5   005 -1.36325911 -1.08660171
#> 6   006 -1.53854617 -0.09012738
#> 7   007 -1.52354538 -0.30411157
#> 8   008  1.15696143  0.93978303
#> 9   009  0.04097976  0.36080386
#> 10  010  0.81092855  0.43480115
#> 11  011 -0.13098113 -0.96315788
#> 12  012 -0.21519443 -1.07255927
#> 13  013 -0.13598839  0.23297971
#> 14  014  0.75063127 -0.17303170
#> 15  015  0.30456751  0.45255576
#> 16  016 -0.17102783 -0.71891022
#> 17  017  0.92218296  0.96628943
#> 18  018 -0.32498667 -0.64252914
#> 19  019 -0.71379643 -1.53126196
#> 20  020  0.05727689 -1.57100199
#> 21  021 -0.87384085 -0.10701433
#> 22  022  0.96971022  0.91163178
#> 23  023  1.03459898  0.87233183
#> 24  024  1.14673880 -0.44672417
#> 25  025  0.43783936  0.28937506
#> 26  026 -1.74606922  0.12186531
#> 27  027 -0.87284557 -0.41133082
#> 28  028  2.29577768  0.20166630
#> 29  029 -0.26535656 -0.57633336
#> 30  030  0.70851604 -0.01306924
#> 31  031  2.25646234  1.10257642
#> 32  032 -0.09872679 -0.59425356
#> 33  033  0.04275459 -0.65130369
#> 34  034 -0.40671575 -1.18693984
#> 35  035  0.04288560  0.11170766
#> 36  036  0.60598403  0.14931716
#> 37  037  1.46050549  0.74471069
#> 38  038  0.23282171 -0.71594193
#> 39  039  0.88042563  1.47032055
#> 40  040  1.52441855  1.77988229
#> 41  041 -1.92760395 -2.46519171
#> 42  042  1.16634023  0.97955929
#> 43  043 -0.39260823 -1.41043447
#> 44  044  1.94880591  0.01694321
#> 45  045 -2.85359956 -1.40198058
#> 46  046 -0.54632149  0.93014893
#> 47  047 -0.16709390  1.04018600
#> 48  048 -1.36056419  0.67210396
#> 49  049  2.97697161  1.63769764
#> 50  050 -0.34985990 -0.66270815
#> 51  051  0.18986435 -1.16230745
#> 52  052  0.27449457 -0.67527848
#> 53  053  0.30286066  0.33793846
#> 54  054 -0.58893406  0.04025576
#> 55  055  0.10365894 -0.26436765
#> 56  056  0.75660319 -0.25990574
#> 57  057 -0.53956006 -1.78772855
#> 58  058  0.89952494 -0.22614466
#> 59  059 -0.19972768  0.26432971
#> 60  060 -0.40193380  1.60375654
#> 61  061 -0.88900196  0.30322454
#> 62  062  0.21553865 -0.15238487
#> 63  063 -0.02307481  0.45599260
#> 64  064  0.02149548  0.66631288
#> 65  065 -0.31549986 -0.47748400
#> 66  066  0.77414399  0.67003098
#> 67  067 -0.34306992  0.09289371
#> 68  068  0.82441957  1.13305860
#> 69  069 -0.87564421 -0.74765277
#> 70  070  0.51119726 -1.25309661
#> 71  071 -1.95935629  1.07202272
#> 72  072  0.16944309  0.82623670
#> 73  073 -0.47560121  0.14664111
#> 74  074  0.20065176  1.80803950
#> 75  075 -0.11067256  0.13076334
#> 76  076 -0.17520246  1.00684264
#> 77  077  0.10163207 -0.12911640
#> 78  078 -0.27201116 -0.51127350
#> 79  079 -1.24710384 -1.30155866
#> 80  080  0.35218419 -0.37386596
#> 81  081 -0.40917018  1.60113122
#> 82  082 -0.53235674  1.57174708
#> 83  083  0.10492315  0.95406557
#> 84  084  0.10754004  0.69245568
#> 85  085 -1.24131337 -1.26610851
#> 86  086 -0.19734099 -0.97189459
#> 87  087 -0.02778819  0.18984383
#> 88  088  0.91638672  0.46690224
#> 89  089  1.00282981  0.38220921
#> 90  090  0.23000886  1.08320098
#> 91  091  0.59603399  0.21702914
#> 92  092 -0.08869654  0.50648246
#> 93  093 -0.13246716  1.02370799
#> 94  094 -0.47241579  0.44424396
#> 95  095  1.55612284  0.60874606
#> 96  096  1.02035843  0.87247995
#> 97  097 -1.05255448  0.11904985
#> 98  098 -0.18267456 -0.35268407
#> 99  099 -2.65669173 -2.26668699
#> 100 100  0.60721146 -0.35650665

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.287780529 -0.12386889
#> 2   002 -0.145174549 -0.89655793
#> 3   003  0.908956136  1.45729901
#> 4   004  1.094638141  0.91297969
#> 5   005  0.461478347 -0.69241874
#> 6   006 -2.221821903  0.05064043
#> 7   007  0.319796437  0.77127133
#> 8   008 -0.368441927 -0.89064137
#> 9   009  0.215858937 -0.26373041
#> 10  010  0.898188419 -0.36764416
#> 11  011 -1.468387604  1.05590260
#> 12  012 -0.287733604 -0.55321357
#> 13  013 -1.110239058 -0.02767260
#> 14  014  0.579970560  0.57002996
#> 15  015 -0.147961470 -0.41082589
#> 16  016  1.070732003  0.29353783
#> 17  017  0.155390679  1.37545503
#> 18  018  0.928204673 -0.40072646
#> 19  019 -1.298770641 -0.82585281
#> 20  020  1.169473134  1.41845147
#> 21  021 -2.655754039 -0.75346500
#> 22  022  0.482794300 -0.68873644
#> 23  023 -1.868078888  2.40131350
#> 24  024  0.227935030  0.03032360
#> 25  025  0.960176925  1.06044110
#> 26  026  0.453742136  0.80736934
#> 27  027 -0.322921117  0.21684189
#> 28  028 -0.245533169 -0.30512636
#> 29  029 -0.636029273  0.73053523
#> 30  030 -0.559625797  0.70978710
#> 31  031 -0.632839594 -0.68071361
#> 32  032  0.844801144  0.78903296
#> 33  033 -1.040438471 -1.42906710
#> 34  034 -0.893685172  0.28411562
#> 35  035  0.534143776  1.71493315
#> 36  036 -1.478183779 -0.85867036
#> 37  037  0.275202321 -0.34931649
#> 38  038  1.427652424 -0.78186229
#> 39  039 -0.050848060 -1.17762903
#> 40  040  0.496432627 -0.78530083
#> 41  041  1.361386692  3.08336804
#> 42  042 -1.022934149 -2.66113389
#> 43  043  0.606043716  0.88044254
#> 44  044  1.856574739  1.86382437
#> 45  045 -0.929356289  0.97764492
#> 46  046 -1.272607597  0.18791825
#> 47  047  0.543922791 -0.26534814
#> 48  048  0.002576702 -1.00083986
#> 49  049 -0.182185775 -0.32833414
#> 50  050  0.404614541 -0.01609144
#> 51  051  1.316630668  1.40000042
#> 52  052 -0.338929515 -0.75394087
#> 53  053  0.628113645 -0.50334200
#> 54  054  1.982643732  0.31771741
#> 55  055  0.336737218 -1.49198672
#> 56  056 -0.271495582 -0.32130617
#> 57  057  0.740382121 -1.49081043
#> 58  058 -1.201410037  0.06195328
#> 59  059  0.547512098  1.49772405
#> 60  060 -0.334518356 -0.28091021
#> 61  061  0.194130594 -0.13920886
#> 62  062 -0.057264955  0.55376788
#> 63  063  0.093073423  0.37894174
#> 64  064  0.107201352  0.66195783
#> 65  065 -1.400010235  0.06564197
#> 66  066 -0.110747950  1.06526751
#> 67  067 -0.138314051 -0.54963996
#> 68  068 -0.044140947  0.08240044
#> 69  069  0.689985694  0.52401981
#> 70  070  0.041555212  0.38976172
#> 71  071  0.583925550 -0.06486424
#> 72  072  0.230821409 -0.79386350
#> 73  073 -1.998489443 -0.38796160
#> 74  074  0.651620841  0.08731138
#> 75  075 -0.163568865 -1.45313490
#> 76  076 -0.286325119 -0.85493973
#> 77  077  0.258709554  0.55070886
#> 78  078 -1.394872584 -1.03608979
#> 79  079  0.593869956  1.43226887
#> 80  080 -0.430132367 -0.45707234
#> 81  081  1.932240534  1.62132722
#> 82  082 -0.185602430  0.80168131
#> 83  083 -1.447880803 -1.33944573
#> 84  084  1.947968071  2.95600301
#> 85  085 -1.485505097 -0.95595458
#> 86  086  0.540047615  0.33799197
#> 87  087  1.227197145  0.16224895
#> 88  088  0.650266310  1.32989687
#> 89  089  1.332927687  1.70477825
#> 90  090  2.138978763  0.28499229
#> 91  091 -0.208464954 -1.62326090
#> 92  092 -0.510355422  1.04809254
#> 93  093 -0.514269020  0.49088427
#> 94  094 -0.491609064 -0.17269315
#> 95  095  0.196629944  0.04603803
#> 96  096  0.233794560 -0.09783641
#> 97  097 -0.805811473 -1.44244333
#> 98  098  0.640495622  0.40583379
#> 99  099 -0.719965139 -0.35534193
#> 100 100  0.114754113  0.55024967