Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5300651
#> Y_2 0.5300651 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1          Y_2
#> 1  0.6583501  0.660103047
#> 2  0.3065682  0.549067596
#> 3 -0.9603424 -1.381082588
#> 4 -0.3997767  0.268009855
#> 5  0.3267301 -0.008166761

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.87204393 -0.805781777
#> 2   002 -0.15531801 -1.534508860
#> 3   003 -0.54794799 -0.642816506
#> 4   004  1.69110391  1.168082034
#> 5   005 -0.83715816 -0.314606157
#> 6   006  1.40616104  1.857652363
#> 7   007 -0.48946969  0.823408162
#> 8   008  0.16213422  1.383281235
#> 9   009  0.64210109  1.539815685
#> 10  010 -1.94953610 -1.560735913
#> 11  011 -1.36724670 -1.326243046
#> 12  012  0.81959716  0.190199379
#> 13  013  1.11444501  2.109491055
#> 14  014 -0.13749337 -0.476830120
#> 15  015 -0.99700177 -0.793760316
#> 16  016 -0.34166640  0.006313479
#> 17  017  0.71902852  0.082406088
#> 18  018 -0.06247588 -0.953356369
#> 19  019 -0.13891797  0.223069942
#> 20  020  1.54057140  0.277222615
#> 21  021  2.86675159  2.981846950
#> 22  022 -1.88823549  0.327747067
#> 23  023 -1.69481791 -1.136476986
#> 24  024 -0.15050341 -0.535327615
#> 25  025  1.16860011  0.861513212
#> 26  026  0.28656176 -0.667937955
#> 27  027  0.42316955  0.842198367
#> 28  028  0.71341643  1.369579953
#> 29  029  0.02777605 -0.427841542
#> 30  030  0.47142181 -0.110053093
#> 31  031  0.44963389  0.884066119
#> 32  032 -0.61008251 -1.395845966
#> 33  033  1.51620956  0.683028010
#> 34  034  2.03443737  1.242803158
#> 35  035 -0.93473071 -0.246902807
#> 36  036  0.32103001  0.120201814
#> 37  037 -0.15982226  0.863925884
#> 38  038  1.52472648  0.929612845
#> 39  039 -0.52968967  0.591219600
#> 40  040 -1.82422517 -2.004825295
#> 41  041  1.51705134 -0.048378119
#> 42  042  0.25793129 -0.400854820
#> 43  043  0.17924478 -1.549953289
#> 44  044  1.19779707 -0.115551096
#> 45  045 -0.29017081  0.910778503
#> 46  046 -0.35674179  0.570208832
#> 47  047  0.93447880  0.233724698
#> 48  048  0.91023253  0.665065859
#> 49  049 -0.96678318  0.899297746
#> 50  050  0.04239940  0.795458596
#> 51  051 -0.00738831  0.708051132
#> 52  052  0.91718796 -0.289478460
#> 53  053 -1.71893007 -1.527501750
#> 54  054  0.09831910  1.061079808
#> 55  055 -0.06691237  0.124916923
#> 56  056  1.13391453  1.441462682
#> 57  057 -1.06676543  0.141420093
#> 58  058 -0.60301824 -0.571785316
#> 59  059  0.23783737  0.266194030
#> 60  060 -0.45520850 -0.845504536
#> 61  061 -2.14039786 -1.162127367
#> 62  062 -0.38137825 -0.747093968
#> 63  063 -0.01711552  0.085450616
#> 64  064 -1.69220194  0.064423685
#> 65  065 -0.54375933  0.095976724
#> 66  066 -1.35549954 -0.895013414
#> 67  067  0.26877200 -0.512665753
#> 68  068  0.26874538  0.130620280
#> 69  069 -0.26719630 -1.097768530
#> 70  070  1.58019752  0.559686878
#> 71  071  0.86916815  1.079809353
#> 72  072 -0.92429021  0.780379538
#> 73  073 -1.78501871 -1.181899010
#> 74  074  0.22437305 -0.719403170
#> 75  075  0.12943645 -0.090060548
#> 76  076  0.44148619 -0.674623215
#> 77  077  0.21276335 -0.063767374
#> 78  078 -0.03015074 -0.097006969
#> 79  079 -0.66385868 -1.142971263
#> 80  080  0.29529589  0.009331799
#> 81  081 -2.12803793 -0.856550750
#> 82  082 -1.32961422 -1.988374486
#> 83  083  0.99808321 -0.776535523
#> 84  084  1.05006994  0.113773338
#> 85  085  0.02188384  0.088277431
#> 86  086 -1.11823188  0.125851032
#> 87  087  0.62766056  1.048953903
#> 88  088 -0.11254129  0.724297840
#> 89  089  0.25597087  0.300324364
#> 90  090 -0.14946980 -0.051168889
#> 91  091  0.79375067  1.459926857
#> 92  092  0.07541825  0.752751287
#> 93  093 -0.63088391 -1.190336442
#> 94  094  0.29606432 -1.211066339
#> 95  095 -0.81532148  0.161995155
#> 96  096  0.20082987 -0.769468702
#> 97  097 -0.27697288 -0.225672607
#> 98  098 -0.77750496 -1.047695726
#> 99  099  0.74531303  0.143959783
#> 100 100 -0.42867358 -0.047067765

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  1.43989915  0.32636743
#> 2   002  1.19828200 -1.24451267
#> 3   003  0.33622641  2.25325159
#> 4   004  0.09547682  1.18590084
#> 5   005 -1.19958193 -1.99298624
#> 6   006  0.11568244  0.81463173
#> 7   007  0.97391271  1.05367410
#> 8   008  0.74616026  0.82692645
#> 9   009 -1.17681061 -0.89904884
#> 10  010  1.04990257  0.28670610
#> 11  011 -0.24729060 -0.83744620
#> 12  012  1.22970849  0.34202779
#> 13  013 -0.60691776 -1.37069304
#> 14  014 -0.26666317 -0.86415600
#> 15  015  0.65602481 -0.31521545
#> 16  016 -0.28313620  0.25911177
#> 17  017 -0.83346165 -0.97725484
#> 18  018 -1.20276845  0.34399254
#> 19  019  1.10566061  0.11100849
#> 20  020 -0.91390282 -0.30230738
#> 21  021 -1.94023171 -0.23480625
#> 22  022  0.09401905  0.04088443
#> 23  023 -0.47727961 -0.10672726
#> 24  024 -0.68961292 -0.83339244
#> 25  025 -0.61076404 -1.65129605
#> 26  026  0.16402004 -1.38089505
#> 27  027 -0.32914409  0.06793411
#> 28  028  1.55470827  0.62564724
#> 29  029  0.80992287  0.17462091
#> 30  030 -0.35390238 -0.80529188
#> 31  031  0.56075049  0.38020240
#> 32  032  0.85843149  0.07276214
#> 33  033  1.20586094  1.32355580
#> 34  034  0.29658974 -0.61038789
#> 35  035  1.38067663  0.07913906
#> 36  036 -0.53550322 -0.71377611
#> 37  037  0.48473142  1.13813276
#> 38  038 -0.21982793 -0.95455628
#> 39  039 -0.35212415  0.31744180
#> 40  040 -1.59614661  0.03647083
#> 41  041  1.58742299 -0.71135798
#> 42  042  1.01525185  0.08599756
#> 43  043 -0.72461603  1.22764357
#> 44  044  0.76249155  0.01409600
#> 45  045 -0.29438972 -1.69546578
#> 46  046  1.37520173  1.80485254
#> 47  047 -0.67294011 -1.67131737
#> 48  048  0.24844096  1.23724824
#> 49  049  0.94952032  1.09872612
#> 50  050  0.55203182  1.09427454
#> 51  051 -0.60323029 -2.09889290
#> 52  052  0.20797934  0.78502630
#> 53  053 -0.33173762 -0.00958902
#> 54  054  1.15895127  0.03909370
#> 55  055 -0.59946626 -1.00041351
#> 56  056  0.71378670  0.10628335
#> 57  057 -0.03752876  1.65594642
#> 58  058  1.03470954  0.73492442
#> 59  059 -1.31521961 -1.05267546
#> 60  060 -1.30332140 -1.72781929
#> 61  061 -0.32949879  0.56432716
#> 62  062 -1.60802823 -0.92124180
#> 63  063 -0.18483163 -0.82797791
#> 64  064  0.48456719  0.71144020
#> 65  065  1.90019565 -0.57124029
#> 66  066  1.23012826 -0.90655013
#> 67  067  0.50983516 -1.73884500
#> 68  068  0.37854065  0.89042800
#> 69  069  0.04014463 -0.04276682
#> 70  070 -2.01787653 -1.54180446
#> 71  071  1.54674592  1.63894884
#> 72  072  0.95509693 -0.46292421
#> 73  073  0.32690278 -1.12578996
#> 74  074 -0.86274555 -0.81345084
#> 75  075 -1.27383438 -0.33610582
#> 76  076 -0.10081584 -0.41502591
#> 77  077  0.56767431 -0.22837784
#> 78  078 -0.63447970 -0.57462367
#> 79  079  1.18898440 -0.21851357
#> 80  080 -0.62748948 -0.42546511
#> 81  081  0.63556221  0.48319537
#> 82  082  0.04108124 -0.09824803
#> 83  083 -0.70841119 -0.72862317
#> 84  084 -0.27782695  0.72254664
#> 85  085 -0.47454152 -0.62148330
#> 86  086 -1.00908816  0.24118016
#> 87  087  0.21852867 -1.15289142
#> 88  088 -0.27754879 -0.59433646
#> 89  089 -0.45795896 -0.25320185
#> 90  090  0.14497883  0.14033565
#> 91  091  0.13204362 -0.06777488
#> 92  092 -0.02163036  1.06581799
#> 93  093  0.29246076  2.12179839
#> 94  094 -0.54932930 -0.41093393
#> 95  095  0.14843374  1.36149567
#> 96  096 -0.88362283 -0.65582815
#> 97  097 -0.24264807  0.62386566
#> 98  098 -0.55887193 -0.57147983
#> 99  099 -1.84539206 -0.47898897
#> 100 100 -1.14194703 -1.47629753