Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4919011
#> Y_2 0.4919011 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.2546133 -0.58665102
#> 2 -0.7710012 -0.37190922
#> 3  2.1239122  1.21922934
#> 4  0.8929818 -0.04739692
#> 5  0.1109488 -0.61297085

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.138410247 -0.194050096
#> 2   002  0.225273332  1.298524681
#> 3   003  1.932914475  0.576479792
#> 4   004 -0.114805959 -0.175123748
#> 5   005 -1.767967593  0.592210863
#> 6   006 -0.333118951 -0.737369361
#> 7   007 -0.266073346 -0.090388979
#> 8   008  0.439555915  1.685105644
#> 9   009  0.927296794  0.935202475
#> 10  010  2.671796055  0.749203443
#> 11  011  0.345606366  0.467170323
#> 12  012 -1.677836617 -0.102837480
#> 13  013 -0.269383653 -0.519888665
#> 14  014 -0.256547401  0.815481692
#> 15  015 -0.507685401 -0.519134307
#> 16  016  0.264068418  0.068819371
#> 17  017  0.125565759 -0.631996798
#> 18  018  0.248159000 -0.171049981
#> 19  019  1.349101466  1.267806600
#> 20  020 -0.541203242 -0.208730702
#> 21  021  0.537378499  0.004218224
#> 22  022 -1.826441444 -0.162782484
#> 23  023 -0.135922316  0.075920563
#> 24  024 -0.918182343 -0.889408903
#> 25  025  0.002372453 -1.069071495
#> 26  026  0.416269207 -0.216466218
#> 27  027  0.641225142 -0.486622222
#> 28  028  1.482217611  0.366110498
#> 29  029 -2.567391941 -1.193370592
#> 30  030  0.143347792  2.253417017
#> 31  031  0.542007237 -0.431676665
#> 32  032 -0.748917243 -0.507736414
#> 33  033 -0.575338563 -0.027517244
#> 34  034  1.013447701  0.164413023
#> 35  035 -0.143768737 -0.485212916
#> 36  036  0.875475846 -0.128623160
#> 37  037 -0.772233271  1.736015717
#> 38  038 -0.518953414 -1.011200130
#> 39  039  1.654808843  0.741009191
#> 40  040 -0.946510042 -0.938383519
#> 41  041 -0.103917153  0.601247559
#> 42  042 -0.190492769 -0.026620135
#> 43  043  1.089864402 -0.237686255
#> 44  044 -0.046427789 -0.250523807
#> 45  045  0.509439769  1.039444625
#> 46  046  0.012680532  0.450490984
#> 47  047 -0.679522625  1.104990435
#> 48  048  1.070111645  1.869605303
#> 49  049  0.129411313  0.707629816
#> 50  050 -1.434683604 -1.916476352
#> 51  051  1.198509457 -0.228968673
#> 52  052  1.410977918 -0.846347838
#> 53  053  0.788805253 -0.207628817
#> 54  054 -2.114757938 -0.861616206
#> 55  055 -1.420082649 -1.058478934
#> 56  056  1.542096306  0.136960122
#> 57  057  1.388788759  0.922843489
#> 58  058  0.462803191 -0.892262053
#> 59  059 -1.208561651  0.201679407
#> 60  060  1.428684017 -0.630929354
#> 61  061 -0.004138671  1.537403440
#> 62  062 -1.413793923 -0.454916605
#> 63  063  1.206742457  1.684211058
#> 64  064  0.117222509  0.081035789
#> 65  065  0.664627587  0.359785074
#> 66  066  0.342060045 -0.257553897
#> 67  067  1.466411781  1.937547683
#> 68  068  0.081562990  1.140359258
#> 69  069  0.697242420  0.775020237
#> 70  070 -0.431008506  1.229461716
#> 71  071  2.640488731  0.615232597
#> 72  072 -1.120070586 -0.743422663
#> 73  073 -2.534030688 -1.396252509
#> 74  074 -1.204201699 -2.266525025
#> 75  075 -0.367364551 -1.260883153
#> 76  076 -0.358060296 -2.217087067
#> 77  077  0.242350024  0.536981620
#> 78  078  1.016549840 -0.287347903
#> 79  079 -0.544267598 -0.096197165
#> 80  080  1.312925101 -0.469778738
#> 81  081  1.197657954  1.494205061
#> 82  082 -1.063176173 -0.298947546
#> 83  083  0.572588830  0.148028463
#> 84  084 -1.322720900 -1.738700723
#> 85  085  0.502240064 -0.321495861
#> 86  086 -0.474842786 -1.066098899
#> 87  087 -0.300128971  0.058696275
#> 88  088 -0.963944745 -0.380642737
#> 89  089  0.562754290 -0.066582890
#> 90  090 -0.276379939  0.125905756
#> 91  091  1.539214840  1.054276067
#> 92  092  0.511063684 -0.220412833
#> 93  093  0.136026958 -0.028724307
#> 94  094  1.073540738  1.321532374
#> 95  095 -0.373113354  0.196763504
#> 96  096 -1.302646578 -0.884771322
#> 97  097  0.650309011  0.178823318
#> 98  098  0.935384366  0.001219096
#> 99  099 -1.887700363 -1.265113979
#> 100 100  0.408593782  0.044953441

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1          Y.2
#> 1   001  1.588620091  0.913263976
#> 2   002 -0.531089257 -1.130167149
#> 3   003  0.940320443  0.756166535
#> 4   004 -1.869438908 -1.050092655
#> 5   005 -1.217261752  0.035081264
#> 6   006 -0.229956828 -1.070591824
#> 7   007  0.612009526  0.089764427
#> 8   008  0.295356972  0.750735511
#> 9   009 -1.792738878 -1.824903825
#> 10  010 -0.677762521 -1.137849161
#> 11  011 -1.060042980 -0.188139051
#> 12  012  1.752682725  1.092355813
#> 13  013  2.579443429  0.822140309
#> 14  014  0.582511208  0.018661487
#> 15  015 -1.342240077 -0.516315222
#> 16  016  0.120612034 -0.932765147
#> 17  017 -0.855799964 -0.764115400
#> 18  018  0.597459673  1.009968830
#> 19  019 -0.566205356 -0.936281152
#> 20  020 -0.821933812 -0.455606815
#> 21  021  1.047815361  1.403223917
#> 22  022 -0.251502790  0.156098466
#> 23  023 -0.382636224 -0.251524901
#> 24  024 -0.009531252  0.088221489
#> 25  025  0.443220195  0.822780362
#> 26  026 -1.119753898 -0.911490108
#> 27  027 -0.885725368 -1.508591715
#> 28  028  0.429661987 -0.661634816
#> 29  029 -1.050589791 -1.178606082
#> 30  030 -1.153175649 -0.903736690
#> 31  031  0.826904324 -0.362006250
#> 32  032 -0.693293304 -0.065812683
#> 33  033  0.047788454 -2.137359521
#> 34  034 -1.376734138 -1.647590966
#> 35  035  1.084264948  2.121478255
#> 36  036  1.483104349  0.256036517
#> 37  037 -0.990449303  0.913111622
#> 38  038  0.415598114  0.109977778
#> 39  039  0.544213991 -0.416823130
#> 40  040 -0.335259344 -1.210324041
#> 41  041 -0.558449521 -0.994809969
#> 42  042 -0.597335498 -1.010953775
#> 43  043 -1.776953516 -2.372067741
#> 44  044  2.268971968  0.831133734
#> 45  045 -0.238797977 -0.151507340
#> 46  046  1.905763666 -0.147711044
#> 47  047 -0.445973508  0.002177668
#> 48  048 -0.898627446 -1.326947761
#> 49  049  1.624434438  1.998958071
#> 50  050  1.865885010  1.708576847
#> 51  051  0.714952817  1.306767247
#> 52  052  3.246204266  0.051831065
#> 53  053  0.999531842  0.104531039
#> 54  054 -1.176222219 -1.263241164
#> 55  055 -2.382904354 -2.603850191
#> 56  056  2.076229081  2.696442673
#> 57  057 -0.147709310  1.097475695
#> 58  058  1.301026233  0.177689512
#> 59  059 -0.056350233  0.573051517
#> 60  060  0.238477030  0.255494820
#> 61  061  0.108557709 -0.059319794
#> 62  062  0.961171106 -0.725183205
#> 63  063  0.152119968 -0.631191599
#> 64  064 -0.672111914 -0.669267460
#> 65  065 -0.214694762 -1.461816104
#> 66  066 -0.809932391 -0.541337275
#> 67  067  1.535881667  0.064079752
#> 68  068  1.156621630  0.469050467
#> 69  069  1.528530684 -1.402262884
#> 70  070 -1.404373505 -1.101287860
#> 71  071  0.515029585 -0.078818721
#> 72  072 -1.263824052 -2.048044151
#> 73  073  0.573428302  0.175608865
#> 74  074 -0.455701293 -0.583566526
#> 75  075  0.818411100  1.223018138
#> 76  076 -0.373003371 -1.538171267
#> 77  077 -1.327283496 -0.492074804
#> 78  078 -0.048364461  0.010978536
#> 79  079  1.345680150  0.052184075
#> 80  080 -0.616325809  0.223157012
#> 81  081  0.124985373 -0.281650331
#> 82  082  0.051474623  1.014071572
#> 83  083 -0.136043507 -0.139169038
#> 84  084 -1.042349425 -0.728542912
#> 85  085  2.105361930  2.521732256
#> 86  086  0.719118648  0.847109696
#> 87  087 -0.892553113 -2.198561099
#> 88  088  1.280713125  0.898309847
#> 89  089  2.249979959 -0.284024404
#> 90  090  0.362002166 -0.328839551
#> 91  091  0.385036342  0.180192913
#> 92  092 -1.285554948 -0.416446353
#> 93  093 -0.437891569  0.163286237
#> 94  094  2.222765181  0.234896581
#> 95  095 -0.756203346 -0.689826556
#> 96  096  0.758614705  1.164509715
#> 97  097  1.614424219  1.208308461
#> 98  098 -0.478458548  0.455506922
#> 99  099 -0.472516836  0.259513726
#> 100 100  0.812852570 -0.227802227