Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4898048
#> Y_2 0.4898048 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  0.2862166  0.66760753
#> 2  0.1756472  0.80376081
#> 3 -0.2389024 -0.60410560
#> 4  1.1316084  0.06070031
#> 5  1.1341388 -0.10378103

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  1.37082323  1.47249798
#> 2   002 -0.29888028  0.31360425
#> 3   003 -0.26917811 -0.10911881
#> 4   004 -0.03424803  0.95106198
#> 5   005  0.54004374 -0.11417907
#> 6   006 -0.68072549 -1.03975424
#> 7   007  0.49113415 -1.01473941
#> 8   008 -0.11316087 -0.71628260
#> 9   009  1.39157472  0.30097058
#> 10  010 -0.45169873 -0.75333051
#> 11  011 -0.18260200 -0.55158810
#> 12  012 -0.11167795  0.08628228
#> 13  013 -1.32155594 -0.39476336
#> 14  014  1.25324937  1.22253240
#> 15  015  0.08215910  0.85092988
#> 16  016 -0.74105542  0.07810046
#> 17  017  1.94989346  1.22417567
#> 18  018 -0.82421198 -0.10097209
#> 19  019 -0.59690490 -1.25852452
#> 20  020  0.49044466 -0.81602427
#> 21  021  0.79780628  1.82005412
#> 22  022  0.01523591  1.49316553
#> 23  023  0.47813836 -0.29617161
#> 24  024  0.33329348  0.28077618
#> 25  025 -1.40562347 -1.17640103
#> 26  026  1.30094103  0.25662979
#> 27  027  1.35848259  0.32595457
#> 28  028 -0.92463586 -1.53115708
#> 29  029  0.06805136  0.02707090
#> 30  030 -0.54995518 -0.30239133
#> 31  031  1.91766278  1.68766456
#> 32  032  1.20349185  0.74121619
#> 33  033 -0.12789645 -0.53362368
#> 34  034 -1.10247673 -1.05594790
#> 35  035 -0.07422361  0.26127015
#> 36  036 -1.35304692 -0.06906265
#> 37  037 -0.14045009  0.37878664
#> 38  038 -1.08552205 -1.11774361
#> 39  039 -0.56087436 -0.52168172
#> 40  040 -1.00365615 -1.80479858
#> 41  041  0.06177141  0.43668394
#> 42  042  0.24799775 -0.22559183
#> 43  043 -0.05098124  0.76696179
#> 44  044 -0.99104467  0.40123630
#> 45  045  0.57005834  1.14227278
#> 46  046 -0.11034228  0.89536314
#> 47  047  0.27668956  2.45583181
#> 48  048 -0.61980452  0.19110322
#> 49  049  0.11666482  0.34836161
#> 50  050  0.66476424 -0.44533224
#> 51  051 -0.59227234  1.78877578
#> 52  052 -0.63721745 -0.25890094
#> 53  053 -0.27489206 -0.16606353
#> 54  054  0.82542993  0.88452304
#> 55  055 -0.18022317 -0.13182660
#> 56  056 -0.23815506  0.49170943
#> 57  057  0.64459678  0.28014538
#> 58  058  0.42511102  0.01607413
#> 59  059 -0.01198131 -0.70819712
#> 60  060 -1.00873181 -0.83416551
#> 61  061 -0.45441152 -0.74095673
#> 62  062 -1.19805149 -1.81858235
#> 63  063  0.58485481  0.17010123
#> 64  064 -0.12243627  0.20408046
#> 65  065 -1.00936372 -0.15810261
#> 66  066 -1.47532003 -0.82249530
#> 67  067  0.86400952  2.54372521
#> 68  068  0.04414711  0.07903668
#> 69  069  1.04931265 -0.44300490
#> 70  070  0.83695347  0.50653118
#> 71  071 -0.88477379 -0.83951603
#> 72  072 -0.93845302 -0.22317316
#> 73  073  0.77233759  1.49875505
#> 74  074 -2.37409475 -1.70002280
#> 75  075  1.37535858 -1.41855218
#> 76  076  1.17419341 -0.11308127
#> 77  077  0.97148624  0.37121448
#> 78  078  1.56863156  1.43256796
#> 79  079 -1.61869815 -1.66399407
#> 80  080 -1.34772495 -0.70811052
#> 81  081 -0.84472808 -1.07311469
#> 82  082 -0.63057427  0.86808967
#> 83  083 -0.12472572  0.63069885
#> 84  084 -0.45680340  0.44757993
#> 85  085 -1.69346957  0.55498395
#> 86  086  0.15736055 -0.21378747
#> 87  087  0.03475985 -0.19083647
#> 88  088 -0.45458870 -0.42728931
#> 89  089 -0.23728638 -1.71344128
#> 90  090  0.42768908  0.98970906
#> 91  091 -1.25105129 -0.58674627
#> 92  092  0.71681131  0.64540734
#> 93  093  0.50011094  0.96304608
#> 94  094  0.32386148 -0.70363846
#> 95  095  0.24303441 -1.22174613
#> 96  096  0.02042723  0.57624933
#> 97  097  0.38969156  0.19501073
#> 98  098  1.57603980 -0.51668274
#> 99  099  0.24472803 -1.40429243
#> 100 100 -1.84013098  0.45763884

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001 -0.3176277024 -0.86551364
#> 2   002  0.8923811248  2.23981041
#> 3   003  1.1139745604  0.78569731
#> 4   004 -0.8670494990 -0.54114310
#> 5   005 -0.5742755906  0.05547689
#> 6   006 -1.3098569132 -1.49015275
#> 7   007  0.9896717131  1.75844672
#> 8   008  1.7032073149  0.53534051
#> 9   009 -2.1525852519 -1.42095663
#> 10  010  0.4569168408  0.50734617
#> 11  011  0.0018953992 -1.04033491
#> 12  012 -1.1117848371  0.10305434
#> 13  013  0.0005280076 -1.03004966
#> 14  014 -1.7805952631  0.57451951
#> 15  015  1.5571985482  1.12130250
#> 16  016 -0.9579112069 -0.93743827
#> 17  017  1.1467793922 -0.92625496
#> 18  018  0.1707389340  0.89412217
#> 19  019  0.1572172067 -0.45687621
#> 20  020 -0.1313949544  0.24258020
#> 21  021 -0.1311368187 -0.02314733
#> 22  022  1.2380534858  0.73396869
#> 23  023  1.2721014719  0.98983676
#> 24  024 -0.6832130200 -0.92308065
#> 25  025 -0.2388508720  0.41602487
#> 26  026  0.0485935095  1.45944674
#> 27  027  0.9838909953  0.03921949
#> 28  028 -0.6842067860 -0.53724265
#> 29  029  0.4349791843 -0.56638699
#> 30  030 -0.2481082066  0.35085811
#> 31  031 -1.4404836693 -0.42334402
#> 32  032 -0.7385839375 -0.65958188
#> 33  033 -1.6545262299 -0.90698324
#> 34  034  1.7986351088  0.40912697
#> 35  035  0.8324031536  0.95573294
#> 36  036  0.3113101797 -0.66253892
#> 37  037 -3.6281924211 -2.41293568
#> 38  038  0.1853761888 -0.80493607
#> 39  039  0.4570783367  0.06353905
#> 40  040  0.0070873142  1.02773732
#> 41  041  0.4244423522  0.94270482
#> 42  042  1.0845967133 -0.15035443
#> 43  043  1.2828912777  0.63835162
#> 44  044 -1.4925980909 -2.51411676
#> 45  045  1.3252768839  0.41409158
#> 46  046 -0.3468509813  0.65916130
#> 47  047  0.9736639214  0.23297667
#> 48  048  1.0412957246  1.27029431
#> 49  049 -0.2589767240 -0.62063824
#> 50  050  0.6750654567  0.58972210
#> 51  051  1.1297821930  0.94933794
#> 52  052 -1.1438127229 -1.16103981
#> 53  053 -2.0091045178 -2.32355020
#> 54  054 -0.3629029766  0.67580797
#> 55  055 -1.4043348770  0.10288065
#> 56  056  1.0207730579 -0.19255980
#> 57  057 -0.4760282321  0.09847368
#> 58  058  1.2137005633  0.63212437
#> 59  059 -0.1445769311  0.96493417
#> 60  060 -0.8514442957 -1.61417796
#> 61  061  0.1640871809  1.73573586
#> 62  062 -1.4294860464 -0.67240432
#> 63  063  0.5764777602 -0.08643461
#> 64  064  1.7005465508  1.55682020
#> 65  065  0.8209320428  0.67370032
#> 66  066  0.8573831884 -0.71178915
#> 67  067 -0.4323003443  0.38265235
#> 68  068  0.0477788884 -0.32725255
#> 69  069 -0.0942524624 -1.59568799
#> 70  070 -0.3687314685 -0.47585028
#> 71  071  0.2891420943 -1.11450623
#> 72  072  0.3111041533  0.57518064
#> 73  073  1.3171052910  1.67260395
#> 74  074  0.9163042214  0.46069321
#> 75  075  0.0875921277 -0.41495965
#> 76  076  0.2301524567  1.80952176
#> 77  077 -1.1377214659 -0.49644801
#> 78  078 -1.6383277711 -0.37960364
#> 79  079  0.0250595200  0.08060371
#> 80  080  0.0004777097 -0.78190760
#> 81  081 -1.0229877390  0.37514638
#> 82  082 -2.0765291908 -0.05885384
#> 83  083  0.2199298014 -0.01854526
#> 84  084  0.5357796444  0.98921926
#> 85  085  0.1155360694  2.13204086
#> 86  086  0.3342072252 -0.33453090
#> 87  087  1.0593544045  1.47349089
#> 88  088  1.3666145944  1.63007795
#> 89  089 -0.9448900240 -0.53189652
#> 90  090  1.1210630096  0.13014861
#> 91  091  0.3507782037  0.55358786
#> 92  092 -2.0541399008 -1.03631517
#> 93  093  0.2660997484 -0.55902224
#> 94  094  1.2341000138  0.06974018
#> 95  095  0.0601794167 -0.79829185
#> 96  096  1.1638134859  0.30911693
#> 97  097  0.6784808544 -0.20209305
#> 98  098  1.2095363889  1.02335429
#> 99  099  0.7135246452 -0.01171350
#> 100 100 -0.2097389565  0.40337199