Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4986505
#> Y_2 0.4986505 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.5073363 -0.46052507
#> 2  0.5834361  0.72965526
#> 3 -0.1370076 -0.09176235
#> 4 -1.1153352 -0.68579866
#> 5 -0.4597734  0.28285073

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -1.523036573 -1.182576081
#> 2   002 -1.760806540 -0.898537255
#> 3   003 -1.571757015 -1.842720314
#> 4   004 -1.349138558 -0.138829614
#> 5   005  0.458607200  1.190331742
#> 6   006  0.670179413 -0.288224395
#> 7   007  1.808702442  1.548565781
#> 8   008 -0.275689025  0.548859971
#> 9   009 -1.006730596 -0.649975358
#> 10  010 -0.539607754 -1.065303003
#> 11  011  0.038032593 -1.780927384
#> 12  012 -1.646796317 -1.083359306
#> 13  013  0.828657183  0.390803166
#> 14  014  0.105608994  1.792630916
#> 15  015 -0.001479774 -0.024940024
#> 16  016 -0.844209003 -0.964219252
#> 17  017  1.172091921  1.518609193
#> 18  018 -1.385198027 -1.319610158
#> 19  019  0.383833591  0.098490628
#> 20  020  0.377212374  0.266181127
#> 21  021  0.620812752  0.759362202
#> 22  022 -1.856312771 -1.796545626
#> 23  023 -2.032030398 -1.929636068
#> 24  024 -0.682119426  0.374790722
#> 25  025  0.148442475 -0.009826204
#> 26  026  0.727891598  0.259889448
#> 27  027 -1.157148490 -0.617048497
#> 28  028 -1.593964561 -2.189778869
#> 29  029 -0.174378402 -0.805555629
#> 30  030  0.440372688 -1.262010093
#> 31  031  1.171569805  1.427462685
#> 32  032 -0.925641781 -1.664919120
#> 33  033  0.134397193  0.366279092
#> 34  034  1.005874990  0.282269991
#> 35  035  0.158079074  0.021783630
#> 36  036  0.191587695 -0.536164906
#> 37  037 -1.731074364 -0.371001138
#> 38  038  0.680624839  0.673190770
#> 39  039 -0.547263445  0.178491453
#> 40  040  0.668154722  0.681156320
#> 41  041 -0.468339280 -0.436383708
#> 42  042 -0.002737316 -1.306100120
#> 43  043  1.972967772  1.350539773
#> 44  044 -1.287676362 -0.615277295
#> 45  045  1.166244849  1.065147346
#> 46  046  0.985409453 -0.737880977
#> 47  047  0.943076577  1.369990167
#> 48  048 -0.794998298  0.615505660
#> 49  049  0.947480470  0.678182480
#> 50  050  0.411367999 -1.435850845
#> 51  051  1.386368933  0.269953255
#> 52  052  1.262964652  1.762992325
#> 53  053  0.987002160  1.661280259
#> 54  054  0.728772294  0.703582277
#> 55  055 -3.144391011 -0.851305087
#> 56  056 -1.055262567 -0.999484000
#> 57  057 -0.119214234  0.284012944
#> 58  058 -2.103895703 -1.986927591
#> 59  059 -0.712929403 -1.487059702
#> 60  060  0.387273593 -0.907062811
#> 61  061  0.870770831  2.226758753
#> 62  062  0.710655990  0.731941530
#> 63  063 -0.684362016  0.426609763
#> 64  064 -2.046520063 -1.411847323
#> 65  065 -1.314557139 -3.036959225
#> 66  066  0.249680643 -0.301587623
#> 67  067  0.460438535  0.898613508
#> 68  068 -0.083007529 -0.212012839
#> 69  069  1.808126246 -0.070089806
#> 70  070 -1.658688722 -0.975043158
#> 71  071  1.221466590  1.687796521
#> 72  072  0.469541472 -0.059212483
#> 73  073  0.231161156  0.453896596
#> 74  074 -0.542139795 -0.830180472
#> 75  075  0.738263989  0.666794398
#> 76  076 -0.345828720  1.027120608
#> 77  077  1.069357484  0.368634694
#> 78  078  0.124011208 -0.545490159
#> 79  079 -1.188064180 -0.237231927
#> 80  080 -0.432792364  1.101760295
#> 81  081  0.797542566  0.942428885
#> 82  082 -0.058613378 -0.008830813
#> 83  083  0.583698604 -0.254269737
#> 84  084  0.372195446  0.450209593
#> 85  085 -1.338042264 -1.025966571
#> 86  086  0.713585428  0.844941011
#> 87  087  0.605538208  0.353703484
#> 88  088  0.381229602 -0.097424422
#> 89  089 -2.208072481 -0.227962464
#> 90  090  0.802499724 -0.371723085
#> 91  091 -0.895301497 -0.961656479
#> 92  092  0.662231102  0.670270604
#> 93  093 -0.008211609  1.422056902
#> 94  094  0.306762090  0.702649511
#> 95  095 -0.120934023  0.281493962
#> 96  096  0.844305710  0.986000508
#> 97  097 -0.272089154 -2.124652376
#> 98  098 -0.264184197  0.534704694
#> 99  099  1.044167543  0.913346678
#> 100 100 -0.364910567  0.164073192

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  1.11758543  1.452699205
#> 2   002 -0.09588737 -1.455363597
#> 3   003 -0.79946309 -0.893366355
#> 4   004  0.49605920  0.696407539
#> 5   005  0.65540986 -1.224885791
#> 6   006 -0.22524584  0.056564792
#> 7   007  0.67354684  1.332454363
#> 8   008  0.30671939  0.579265567
#> 9   009  1.27553597  0.404279790
#> 10  010 -2.03456736 -1.999072513
#> 11  011 -0.86952800  0.250390247
#> 12  012 -1.14151607 -0.216400635
#> 13  013  0.76443182  1.535232874
#> 14  014 -0.63626191 -1.449367309
#> 15  015  0.72820552  0.466648278
#> 16  016  0.04948760 -1.042541165
#> 17  017  1.53516526  1.219554526
#> 18  018  0.21235476 -1.455154824
#> 19  019  0.36013161 -0.172298350
#> 20  020  0.57491787  0.287568956
#> 21  021  2.16589612  0.761772491
#> 22  022 -0.29466505  0.155553756
#> 23  023 -0.52841011 -0.650956454
#> 24  024  1.49270544  0.819307353
#> 25  025  0.23341945 -1.246087741
#> 26  026  1.41427762  0.067959299
#> 27  027 -0.22802369 -2.208480597
#> 28  028 -1.71075370 -1.082315471
#> 29  029  0.87357651 -0.683423104
#> 30  030 -0.71853946 -0.235597812
#> 31  031 -1.94070300 -1.699004726
#> 32  032  0.19060424 -1.747391383
#> 33  033 -1.78158453 -2.614748174
#> 34  034  0.91455875  0.595110884
#> 35  035 -0.40079914  0.234040220
#> 36  036  0.76038140  1.804850193
#> 37  037  0.23660908  0.089950468
#> 38  038  1.03657045 -1.214668306
#> 39  039  2.18127621 -0.334390159
#> 40  040 -0.34734423 -1.550800176
#> 41  041 -1.54734196 -0.600364166
#> 42  042  0.61085770 -0.996937614
#> 43  043  0.61259096 -0.319580426
#> 44  044 -1.52122505 -1.345328407
#> 45  045 -1.41956989 -1.059579831
#> 46  046  0.59231887  0.365532016
#> 47  047 -0.08159622 -0.225711805
#> 48  048  0.09486424  0.007926136
#> 49  049  1.56096478  0.548031026
#> 50  050 -0.48808034  0.591250035
#> 51  051  0.74430163  1.445652685
#> 52  052  0.71010858  0.573200575
#> 53  053 -0.65511005  0.575852989
#> 54  054 -0.45041326  0.016364167
#> 55  055  0.52716542 -0.297693251
#> 56  056 -1.73808599 -0.258900561
#> 57  057 -1.41444037 -1.354094561
#> 58  058  0.40869947  0.954805324
#> 59  059 -1.09821088 -0.596695469
#> 60  060  0.85649697  0.144619655
#> 61  061 -1.00267595  1.744752834
#> 62  062 -0.15553251  1.691832070
#> 63  063  0.16895349 -0.341578259
#> 64  064 -0.64085607 -0.457962259
#> 65  065 -0.49247758 -0.411621089
#> 66  066  1.32622573 -0.584248698
#> 67  067 -2.62596985 -1.911320715
#> 68  068 -1.26811314 -0.621166950
#> 69  069 -1.31327280  0.527533662
#> 70  070 -2.29007841 -0.914918477
#> 71  071 -0.47778157 -0.004892902
#> 72  072  0.28897405 -1.195977648
#> 73  073 -0.02407889 -0.214587672
#> 74  074 -0.73009349 -1.149032607
#> 75  075  0.39046905 -1.035990842
#> 76  076 -1.56009186 -1.120860228
#> 77  077  0.68496214 -0.607910167
#> 78  078  0.29121698 -0.045512297
#> 79  079  2.74320700  0.834186087
#> 80  080 -0.99128362 -1.865419504
#> 81  081 -0.66602469  0.418448891
#> 82  082 -0.71592445 -1.527416603
#> 83  083  0.09360577 -0.018534494
#> 84  084 -1.85707599 -1.406799593
#> 85  085 -0.46158803 -0.504509600
#> 86  086 -2.15588525 -1.173862726
#> 87  087  1.17337676 -0.987964806
#> 88  088  0.43812256 -0.279894765
#> 89  089  0.58542560  0.593035348
#> 90  090  0.03459291  1.757082293
#> 91  091  0.83258189  1.305643258
#> 92  092 -0.47196894 -2.103477167
#> 93  093 -1.46850252 -0.158194173
#> 94  094  0.12483143 -0.491727059
#> 95  095  0.02850901 -1.951706775
#> 96  096 -0.62309124 -0.670919046
#> 97  097 -1.11190657 -0.685022027
#> 98  098  0.19806641 -0.713394043
#> 99  099 -0.26997471 -0.635977736
#> 100 100  1.85179936  1.714306050