Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5141007
#> Y_2 0.5141007 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.3125908  1.41273940
#> 2  1.0388606 -0.07352465
#> 3 -0.2069020 -0.50164204
#> 4  0.4146214  0.64825334
#> 5 -1.2561536 -1.10802753

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -0.006675606 -1.57109082
#> 2   002  0.903385752  0.64422540
#> 3   003  1.127801870  0.44917643
#> 4   004 -1.917089658 -0.43980315
#> 5   005 -0.385935672 -0.68310876
#> 6   006  0.919976110 -0.83054962
#> 7   007 -2.342242094 -0.31068812
#> 8   008  0.240326896  0.86291136
#> 9   009 -0.412906568 -0.73718629
#> 10  010 -0.284580257  0.46742610
#> 11  011  0.030859057 -0.71302687
#> 12  012 -0.834481572 -1.40700835
#> 13  013  1.545952721  0.65692999
#> 14  014 -0.595788118 -0.18501350
#> 15  015 -1.300958219  0.22139876
#> 16  016 -1.157857834 -0.44356904
#> 17  017  0.805286380  1.16030293
#> 18  018 -0.941168434  0.96650090
#> 19  019  1.318971662  0.92015796
#> 20  020  1.527536367  2.26682066
#> 21  021 -1.313477840 -0.51715694
#> 22  022  0.321125694  0.19672483
#> 23  023  2.250624800  2.15211271
#> 24  024  1.321434371  1.42405577
#> 25  025  0.601846527 -0.18354460
#> 26  026 -2.412441871 -0.84178603
#> 27  027  0.297391995 -0.01924673
#> 28  028 -1.358918722 -0.87281482
#> 29  029 -0.804163819 -0.24287040
#> 30  030 -1.414654855 -1.73000649
#> 31  031  0.760401530 -1.39919029
#> 32  032 -0.896582327 -1.94185869
#> 33  033  0.362573055  0.20214944
#> 34  034 -0.906441298 -0.91836145
#> 35  035  0.541776617  0.60936285
#> 36  036 -0.668719995  1.90302323
#> 37  037 -0.905528664 -1.41272437
#> 38  038  0.405360625  0.14910773
#> 39  039 -0.426307490 -1.73533143
#> 40  040 -1.564720090 -1.68795996
#> 41  041 -0.472252044  0.66856897
#> 42  042 -0.984285093 -0.36498103
#> 43  043  0.627883956 -0.42415349
#> 44  044 -0.259601848  1.25245840
#> 45  045  1.553311498 -0.43528466
#> 46  046 -0.521127926 -0.77997587
#> 47  047  1.000621924  0.72276611
#> 48  048 -0.059879447 -0.35766793
#> 49  049  0.332598121  0.60668007
#> 50  050  1.737769994  0.95856881
#> 51  051  0.273377952 -0.85630180
#> 52  052  1.476500495  0.26126744
#> 53  053  0.038211512  0.21980454
#> 54  054  0.963407441  0.47710655
#> 55  055 -1.200503643  1.02154314
#> 56  056 -0.300210628 -0.20811547
#> 57  057  1.253201785  1.88917836
#> 58  058  0.787752372 -1.07213167
#> 59  059  0.719529474  0.45103906
#> 60  060 -0.197004688 -0.91540100
#> 61  061 -0.817674659 -1.48991733
#> 62  062 -1.545108039 -0.75649355
#> 63  063  0.900207851  0.36332357
#> 64  064  0.884414839  1.48973092
#> 65  065 -1.391986687 -0.45276338
#> 66  066  1.280726999  1.79136619
#> 67  067 -0.026366084  0.31813284
#> 68  068  1.625721814 -0.37383307
#> 69  069  2.270813579  1.80528340
#> 70  070 -0.588444503 -0.65930669
#> 71  071 -1.070880898  2.28517358
#> 72  072  2.183496983  1.60388129
#> 73  073  0.862315664  0.15938653
#> 74  074  0.309520435  2.33017430
#> 75  075  0.713018142 -0.50056259
#> 76  076 -1.619041761 -0.27248932
#> 77  077 -0.111483834 -0.71644454
#> 78  078 -0.182514828  0.02200448
#> 79  079 -1.229132258  0.08058959
#> 80  080  0.539553491  0.49521874
#> 81  081 -0.461604462 -0.15615108
#> 82  082  0.088841091  2.22814043
#> 83  083  1.699464832  2.28480859
#> 84  084 -0.339402562 -0.34875007
#> 85  085 -0.251929070 -1.19346166
#> 86  086  1.038949610  1.63269070
#> 87  087  1.158588784  1.05744850
#> 88  088 -1.258448665 -0.95886908
#> 89  089  0.888484560  1.29964785
#> 90  090  0.695839189 -0.41751378
#> 91  091 -0.793731403 -1.77179005
#> 92  092  0.020837895  0.96543883
#> 93  093 -0.932180759  0.26793017
#> 94  094 -1.118952588  0.81470739
#> 95  095 -1.077424311  1.26406167
#> 96  096  0.270537717  0.04784194
#> 97  097  1.373472106 -0.38444331
#> 98  098 -1.856551784 -0.66346919
#> 99  099 -0.301021527 -1.85424377
#> 100 100  0.539844139 -0.24304350

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  2.168542695  1.96025833
#> 2   002  0.815534767  1.02883813
#> 3   003 -1.243807570  1.05819822
#> 4   004  1.532668472  0.53522256
#> 5   005 -0.610602464  0.21594527
#> 6   006  0.743206445  1.70454228
#> 7   007  0.534263612  1.10786035
#> 8   008  0.882078120  1.33770498
#> 9   009 -2.385078082 -1.11322052
#> 10  010 -0.253822081  0.86461660
#> 11  011 -1.839338997 -1.30070941
#> 12  012 -0.023649888 -0.12105198
#> 13  013 -0.559618718 -0.69635763
#> 14  014  0.882351997  0.99961402
#> 15  015 -2.035007408 -3.02185657
#> 16  016  0.367271191 -1.28072213
#> 17  017 -0.184347511 -0.25118735
#> 18  018 -1.142867099 -0.68151233
#> 19  019 -0.776838391 -0.07560318
#> 20  020  0.071223688 -0.12434148
#> 21  021 -0.712054573  0.51583359
#> 22  022  2.547635934  1.47634417
#> 23  023  0.317487768  1.10734736
#> 24  024 -0.285779823 -0.32322988
#> 25  025 -0.034000788 -0.12571533
#> 26  026 -0.674384378 -0.20614016
#> 27  027 -0.254260481 -0.14138518
#> 28  028 -1.153251027 -1.26225412
#> 29  029  1.236487719  0.40675851
#> 30  030 -0.487410728  0.51613595
#> 31  031  0.010603004  0.76476157
#> 32  032 -0.189560158 -0.86939935
#> 33  033  0.315022099 -0.69161149
#> 34  034 -0.630597585  0.33572850
#> 35  035 -0.819967593 -0.52533104
#> 36  036  1.226875562  3.00977192
#> 37  037 -0.762373609 -1.98887669
#> 38  038  0.890802631  1.22336984
#> 39  039 -0.468604996  0.47074873
#> 40  040 -0.334271365  0.52622011
#> 41  041  0.050228824 -0.09287324
#> 42  042 -0.023253285 -0.37868946
#> 43  043 -0.593603782 -0.38255177
#> 44  044 -0.531138125 -1.58742623
#> 45  045 -0.260161520 -0.77374200
#> 46  046 -1.265436995 -0.86210301
#> 47  047 -1.956405431 -1.28360943
#> 48  048  0.002642026 -1.39372409
#> 49  049 -0.931919371  1.07665379
#> 50  050  0.958686135  1.64472047
#> 51  051 -0.756069716  0.84079335
#> 52  052  0.315200410 -1.01803263
#> 53  053 -0.178394254  0.16655102
#> 54  054  0.456770455  0.49181420
#> 55  055 -0.264554602  0.10843778
#> 56  056 -0.639661759 -0.37942128
#> 57  057  0.760033173  0.44387073
#> 58  058 -1.496853895 -1.81585779
#> 59  059  0.727379560 -1.01144547
#> 60  060  0.531732949  0.98788879
#> 61  061 -0.500620644 -0.60082693
#> 62  062 -0.636294205 -0.46395245
#> 63  063  0.290805858 -1.02064777
#> 64  064 -1.507641479  0.22135716
#> 65  065  0.929163931  1.07761312
#> 66  066 -1.460650992 -1.57438483
#> 67  067 -0.580961004  0.02361167
#> 68  068  0.492103858  0.07239265
#> 69  069  0.593165800  0.24492142
#> 70  070 -0.605672733  0.44332586
#> 71  071 -0.234107775 -0.44128007
#> 72  072  0.057385379 -0.67797285
#> 73  073 -1.910370759 -0.66153453
#> 74  074 -1.093577396 -1.08794177
#> 75  075 -1.069071257 -1.77346556
#> 76  076  0.388341494  1.48536031
#> 77  077 -1.629215238 -1.06256674
#> 78  078  1.206008210  2.42697896
#> 79  079  0.690172627  0.06397659
#> 80  080 -0.929556081 -0.56595644
#> 81  081  0.480795954  0.17390333
#> 82  082  0.619703308  0.71248958
#> 83  083 -0.312015316 -0.20035872
#> 84  084  0.204853540 -2.09080472
#> 85  085 -1.071120839  0.17220654
#> 86  086  1.086755892 -1.16221464
#> 87  087 -1.154162926  0.99708152
#> 88  088  1.157854463  0.12009985
#> 89  089  0.029577361  0.47675084
#> 90  090 -0.873896750 -1.40525955
#> 91  091  0.949921717  0.44854938
#> 92  092  0.706544661  1.49695362
#> 93  093  1.098247202 -0.91993907
#> 94  094 -0.930401339 -0.48436532
#> 95  095  0.544260041  1.21059639
#> 96  096  0.639783822  1.08257936
#> 97  097 -0.416686524 -0.05892673
#> 98  098 -1.331811803 -1.78659065
#> 99  099 -1.369527258 -1.05725052
#> 100 100 -0.383896564  0.02295819