Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4783733
#> Y_2 0.4783733 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -0.4509253  1.35257959
#> 2  1.2550294  0.86933310
#> 3  0.4572750 -0.04153656
#> 4  0.6707378  0.45301909
#> 5 -0.8374090 -1.14801566

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -1.602343155 -1.14207767
#> 2   002 -0.491774875 -0.12811094
#> 3   003  0.256859888 -0.47575906
#> 4   004  0.384095693  0.61362925
#> 5   005 -0.560080444  1.20513102
#> 6   006  0.265853538  0.45830405
#> 7   007 -2.186240252 -1.57363984
#> 8   008  0.210491466  1.94241690
#> 9   009 -0.314074281  0.16686697
#> 10  010  0.343252761 -0.17518500
#> 11  011  1.058501220  0.96676615
#> 12  012  1.392780375 -0.02243059
#> 13  013 -0.453733671 -0.90672623
#> 14  014  1.529178657 -0.41492669
#> 15  015  0.483085824 -0.09539045
#> 16  016  0.503032405  0.18680025
#> 17  017 -0.415480607 -0.93881499
#> 18  018  0.783890628 -0.04857732
#> 19  019 -0.038445155 -1.30393698
#> 20  020  0.710105285  1.51470027
#> 21  021  0.523428111  0.59972503
#> 22  022 -2.552141265 -1.56153823
#> 23  023  0.372392693  0.66128957
#> 24  024  0.167765619  1.17115713
#> 25  025 -1.921695412 -1.41169606
#> 26  026  2.247418823  0.83772728
#> 27  027  1.044786451 -0.36280367
#> 28  028  0.154039365 -0.86074026
#> 29  029  0.086511423  1.01461637
#> 30  030 -0.937168473 -1.67836173
#> 31  031 -0.433363118 -0.53360913
#> 32  032  0.824391309 -0.40359769
#> 33  033 -0.112125824  1.37827568
#> 34  034  1.160963184  0.67670996
#> 35  035  0.578083123  0.64038898
#> 36  036 -1.418193619 -1.98139069
#> 37  037  1.245933621 -1.70989613
#> 38  038  0.024084704 -0.05741015
#> 39  039 -0.950130697  0.19736628
#> 40  040 -0.234416923 -0.37812187
#> 41  041  1.144272152  1.46741743
#> 42  042  3.136917735  2.26692491
#> 43  043  0.791354975  0.31848445
#> 44  044  0.824072446  0.65328351
#> 45  045 -1.495719792 -0.90967461
#> 46  046 -1.176114767 -1.41428534
#> 47  047  1.409419433  1.47253866
#> 48  048 -0.470730198 -1.82079249
#> 49  049 -1.060068435 -0.79869317
#> 50  050 -1.955039832 -2.20441173
#> 51  051  1.085284525 -0.58982399
#> 52  052 -0.009935763  0.10376054
#> 53  053  0.982176283  0.78695960
#> 54  054 -0.142029703  0.89334839
#> 55  055 -0.473786679 -1.73037237
#> 56  056  2.039069515  0.96562945
#> 57  057  0.456652546  0.47419919
#> 58  058 -0.054791544 -0.39483219
#> 59  059 -1.586102715  0.17318884
#> 60  060 -1.763227481 -1.86897335
#> 61  061 -0.254151818 -1.54471138
#> 62  062  1.672025214  0.07423228
#> 63  063  0.508877851  0.17313638
#> 64  064 -0.514532474  0.75397189
#> 65  065 -0.529728712  0.62252883
#> 66  066  0.285358404 -0.55071138
#> 67  067  0.274127284  0.08763393
#> 68  068  1.177152921 -1.40874626
#> 69  069 -1.534500400 -0.31102293
#> 70  070  0.319262699 -0.30655850
#> 71  071 -1.807328581 -1.65564307
#> 72  072  0.257631716  1.17807635
#> 73  073  2.764209879  0.61064302
#> 74  074 -0.128980308  1.06837828
#> 75  075 -1.327539361 -0.48322548
#> 76  076 -0.150152123 -0.87438147
#> 77  077  2.493983705  1.27419976
#> 78  078  1.420456863  1.02587808
#> 79  079 -0.997578229 -1.35390688
#> 80  080  1.161239840 -1.09018900
#> 81  081  1.297464564  0.51123475
#> 82  082  1.960531562  1.39548037
#> 83  083 -0.440913428  0.28962192
#> 84  084 -0.272073663 -0.65271174
#> 85  085  0.624701324  1.19483456
#> 86  086 -1.850098854 -1.79995745
#> 87  087 -1.797679957 -0.29434691
#> 88  088 -0.700653207 -1.15484070
#> 89  089 -0.877166785  0.26586685
#> 90  090  0.602063952 -1.45083176
#> 91  091 -0.584507840 -2.02103772
#> 92  092 -0.190769849 -0.65488281
#> 93  093  1.225882163  2.05161038
#> 94  094 -1.027857253  0.55291221
#> 95  095  0.008579656  1.12034875
#> 96  096 -0.793541451 -0.14851897
#> 97  097  0.124405742  0.61883698
#> 98  098  0.986397160 -1.72246381
#> 99  099 -0.777628159 -0.78227490
#> 100 100 -1.792135892  0.22632626

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.04316085  1.054387793
#> 2   002  0.10385528  0.609159478
#> 3   003  0.51946999  0.118381407
#> 4   004 -1.07146190  0.334461749
#> 5   005 -0.07055086 -1.181948048
#> 6   006 -1.20357785 -0.348609317
#> 7   007 -1.60825159 -0.837450054
#> 8   008 -0.42209082 -0.026303496
#> 9   009 -0.43369007  0.476335485
#> 10  010  0.16974276 -1.344889187
#> 11  011  0.95645947  1.715340166
#> 12  012 -2.97796059  0.226984552
#> 13  013  0.63056090 -0.246201278
#> 14  014 -0.22460601 -0.055555970
#> 15  015  0.18695322  0.789707781
#> 16  016  0.70785870  0.575124282
#> 17  017 -0.48366142  0.598917211
#> 18  018  1.89664401  0.972276382
#> 19  019 -2.55333906 -2.568006211
#> 20  020  0.46004363  0.506821286
#> 21  021 -0.11612119  1.168360147
#> 22  022 -0.60656023 -2.077477389
#> 23  023 -0.53393209  1.116272574
#> 24  024  0.22235073 -0.465089345
#> 25  025 -0.16761979 -0.860680725
#> 26  026  1.14968960  0.768078757
#> 27  027  0.24180500  1.435779539
#> 28  028 -0.96905314 -0.264278294
#> 29  029 -0.47447901 -0.514429605
#> 30  030 -0.08353890  0.678354795
#> 31  031 -0.37347507  0.185514132
#> 32  032 -2.09696696  0.034722847
#> 33  033  2.19823624  0.887601543
#> 34  034  0.45845984  0.138693099
#> 35  035  1.56017733  1.374877331
#> 36  036 -0.49325656 -0.967476808
#> 37  037  0.25340244  0.825803985
#> 38  038  1.19210844  0.865652505
#> 39  039 -0.90290117 -0.483139412
#> 40  040 -0.63400350 -1.347163737
#> 41  041 -0.37017399 -1.185320866
#> 42  042 -1.48938689 -0.206323504
#> 43  043  1.75829020  0.902013355
#> 44  044  0.07541540  0.111113674
#> 45  045 -0.16967113 -0.519368393
#> 46  046  0.42156244 -0.710604698
#> 47  047 -1.03399348 -0.878460054
#> 48  048  0.75822770  0.169686796
#> 49  049  0.20398357 -1.513940546
#> 50  050  0.57737230  0.720623428
#> 51  051 -0.33216441  1.402282554
#> 52  052 -0.92376430 -0.491720661
#> 53  053 -1.13837755 -1.078222444
#> 54  054 -0.12965106  1.071375068
#> 55  055 -2.20322128 -0.431766550
#> 56  056 -1.04698226 -0.283498268
#> 57  057 -0.18838574 -0.561592397
#> 58  058  0.33695963  1.096252041
#> 59  059 -1.02541121 -0.249525949
#> 60  060 -1.47516076 -0.465271847
#> 61  061  0.86217996  0.415818788
#> 62  062  0.24909685  1.474641373
#> 63  063  0.08825478  0.215385045
#> 64  064 -0.71840296 -1.590414606
#> 65  065 -0.25464086  0.339621679
#> 66  066 -0.56243392 -0.097614271
#> 67  067 -0.02744809 -1.899192276
#> 68  068  0.41687341  1.089837031
#> 69  069  0.11084217  1.120618833
#> 70  070  0.41800227  0.958196985
#> 71  071 -0.31019102 -2.030949848
#> 72  072 -1.24338218 -0.408418610
#> 73  073 -0.51407594 -0.632959047
#> 74  074 -0.28617163  0.052831559
#> 75  075  1.08053434  0.513181691
#> 76  076  0.63934956  0.534575096
#> 77  077  1.76838785  3.069115865
#> 78  078  1.17111348  0.939574971
#> 79  079 -0.54594938 -0.539119510
#> 80  080 -0.58875032  0.382662715
#> 81  081 -0.49938163 -0.681471830
#> 82  082 -0.60081487  0.029532596
#> 83  083 -1.17552552  0.256701594
#> 84  084  0.02099395  0.420567404
#> 85  085 -0.85648909  0.663269971
#> 86  086  0.12086044  0.373629507
#> 87  087 -1.07815881 -2.393841443
#> 88  088 -0.29180505 -0.317252657
#> 89  089 -0.72162910  0.343716809
#> 90  090 -0.73346967 -2.588415567
#> 91  091 -0.03013008 -0.500508197
#> 92  092  1.13248258  0.872230582
#> 93  093  2.27661252  1.452548085
#> 94  094  1.74504072 -0.791463787
#> 95  095  0.31257676  0.116017745
#> 96  096 -0.07236995 -0.008617042
#> 97  097  0.88746817  0.198251844
#> 98  098  0.78276059  0.536862768
#> 99  099  0.66495451 -0.561405143
#> 100 100 -0.84783433 -1.633761874