Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4789956
#> Y_2 0.4789956 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1726039 -1.4039507
#> 2  0.1344726  0.8179799
#> 3 -0.6192327  0.7133114
#> 4  1.2035543  0.7481318
#> 5  0.9863459  0.7160478

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.621630100 -0.860151452
#> 2   002  0.603037302  0.098723647
#> 3   003  1.434798078  0.551968719
#> 4   004  0.434519076  0.320220034
#> 5   005  0.712651569  0.024666998
#> 6   006 -0.105076712 -0.608908967
#> 7   007  1.738969727 -0.026610825
#> 8   008 -0.909717116 -0.082021915
#> 9   009  0.521752350  0.967266137
#> 10  010 -0.755343996 -0.530663709
#> 11  011  0.936194581  0.935388287
#> 12  012  0.144045443  1.115086566
#> 13  013  0.015267977  1.104634661
#> 14  014 -0.102182777 -1.216227984
#> 15  015  0.673171740  0.219983022
#> 16  016  0.624805452 -0.419725821
#> 17  017  0.289146447  0.969113406
#> 18  018  1.184254321  0.014007875
#> 19  019  0.735553034 -0.079599959
#> 20  020  0.231202184  1.111420338
#> 21  021  0.150692298  1.349068731
#> 22  022 -0.052192522  0.990633638
#> 23  023 -0.549089268 -0.356979788
#> 24  024 -0.040542291  1.277912336
#> 25  025  1.134795093  1.539748448
#> 26  026 -0.750910595 -1.367505296
#> 27  027  0.972061305  0.961582043
#> 28  028 -0.232538607  1.096424993
#> 29  029 -1.164389614 -0.118783241
#> 30  030 -1.154678113 -0.636077052
#> 31  031 -0.058006508  0.077760371
#> 32  032  0.652776901 -0.837584453
#> 33  033  0.800838307  0.074648216
#> 34  034  0.209765295  1.030713020
#> 35  035  0.407093961  0.170111993
#> 36  036 -2.134426192 -0.544154335
#> 37  037 -1.292076589 -1.010404721
#> 38  038 -1.427363300 -1.631442302
#> 39  039 -0.015635668  1.118731094
#> 40  040  0.621733548 -0.457975144
#> 41  041  0.060753487  1.523152308
#> 42  042  2.538216752  0.951550202
#> 43  043  1.647528959 -0.687945077
#> 44  044  0.352262686  0.265402324
#> 45  045  0.392628407 -0.125764667
#> 46  046 -0.736215589  1.443638429
#> 47  047 -1.317994285 -1.298890398
#> 48  048 -0.062846706  0.400064319
#> 49  049  1.480850037  1.506341236
#> 50  050  0.116279203  1.520938123
#> 51  051  0.585303486  1.169372851
#> 52  052  0.291730714  0.736652016
#> 53  053  0.322895285 -0.902086969
#> 54  054 -0.770257692 -0.915735852
#> 55  055  0.102204633  0.283521180
#> 56  056 -0.349672353  0.542512039
#> 57  057 -0.657965133 -0.913532132
#> 58  058 -0.404746902 -0.140871101
#> 59  059  1.072925153  1.894430424
#> 60  060 -0.206791738  0.659813343
#> 61  061  1.443672363 -1.017731617
#> 62  062 -0.620181023  0.567455027
#> 63  063 -0.289823672  0.172158844
#> 64  064 -0.210736644  0.204366774
#> 65  065 -2.085913956 -0.807309246
#> 66  066  0.879782834  0.565417513
#> 67  067 -1.518303012 -0.524488041
#> 68  068  0.409728520 -1.694187881
#> 69  069  1.290541937  0.376467639
#> 70  070  1.069787443  0.937316160
#> 71  071  0.631860535 -0.073434531
#> 72  072 -0.812680869  0.111290988
#> 73  073  0.021884337 -0.325080353
#> 74  074  1.219634421  0.286177463
#> 75  075  0.513227757  1.562342757
#> 76  076  0.777586119  1.997557913
#> 77  077  0.499782341 -0.026409377
#> 78  078  1.226202644 -0.589318782
#> 79  079 -0.167033421 -0.086033919
#> 80  080 -0.228156832  1.067009914
#> 81  081  0.654548254  1.571704616
#> 82  082 -0.006636738  1.047527223
#> 83  083  0.151833706 -0.247723820
#> 84  084 -0.071375352  0.070829220
#> 85  085  0.517043949 -0.243670461
#> 86  086  1.596244638  2.162091712
#> 87  087  0.432932855 -0.625491157
#> 88  088  0.740849045 -0.610110936
#> 89  089  0.474637302  0.540670045
#> 90  090 -1.290905464 -0.650417185
#> 91  091  0.308395464  0.493153372
#> 92  092  1.539582161  1.081094769
#> 93  093 -0.459786960 -0.008270616
#> 94  094 -2.066356698 -1.294525028
#> 95  095  0.512542986  0.727346117
#> 96  096 -0.473760704  0.667359618
#> 97  097 -1.132518409 -1.083873864
#> 98  098  1.160091491  1.232177513
#> 99  099  0.216335113 -1.120433950
#> 100 100 -0.177170043 -0.418505235

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1           Y.2
#> 1   001 -0.25101792  0.4320938610
#> 2   002 -1.39808281  0.4570540853
#> 3   003 -0.22889484  0.1042053212
#> 4   004  0.01575307 -0.9370657351
#> 5   005 -1.73457538 -1.3673327658
#> 6   006 -0.93759034 -1.0323774254
#> 7   007  0.65711573  1.6906765913
#> 8   008  1.47958515  2.1761807443
#> 9   009  1.05158355 -0.4934739970
#> 10  010 -0.95776833  0.6591951352
#> 11  011 -0.43127871  0.6533094574
#> 12  012 -0.80408876 -1.0303137276
#> 13  013  1.21232533  0.8509900333
#> 14  014 -1.04358667 -1.2945873661
#> 15  015  0.48173086  0.4281287396
#> 16  016 -1.60102028 -0.1324455108
#> 17  017 -0.51005573  0.0254430928
#> 18  018  0.51528998 -0.5554820711
#> 19  019 -2.36895118 -0.9788953394
#> 20  020  0.09597600  0.3124930793
#> 21  021  0.12226151 -0.6005503713
#> 22  022  0.12883039  1.0760789972
#> 23  023 -0.06976805 -0.7693006372
#> 24  024 -0.90708235 -1.4204891275
#> 25  025  0.85993339  0.8545301678
#> 26  026 -1.74483021 -0.2492734570
#> 27  027  0.28959459  0.4748998557
#> 28  028 -0.67285230 -1.1382597620
#> 29  029  1.00694756 -0.3552966563
#> 30  030  0.07292272  0.1853742490
#> 31  031 -0.34919142  0.4463214418
#> 32  032  0.49548857  0.9920513090
#> 33  033 -1.00685947 -1.8022455606
#> 34  034  0.95435355  0.7819663857
#> 35  035  0.32235244  1.8820534745
#> 36  036  0.63002944 -0.0053445165
#> 37  037  0.42995444 -0.0447588292
#> 38  038 -0.36619014 -0.8478548461
#> 39  039 -1.36897825  0.7226969936
#> 40  040 -1.22757780 -0.7626550388
#> 41  041  0.58281493  1.2376843485
#> 42  042  0.97965841 -0.5640612873
#> 43  043 -1.67832902  0.0587381781
#> 44  044 -0.18197983  0.1900923696
#> 45  045  0.80331360 -0.1822424036
#> 46  046  0.25251671  0.4854125315
#> 47  047 -0.79399369 -0.3985506977
#> 48  048  0.96754969  0.3082250380
#> 49  049  0.44571691 -0.0008130536
#> 50  050  1.76030279 -0.1041847739
#> 51  051 -0.08548798 -0.7051132023
#> 52  052  0.19835837 -0.5782878475
#> 53  053  0.30662217 -0.8244869199
#> 54  054  0.71421639  1.2253193885
#> 55  055  0.57889635 -0.3103681092
#> 56  056  0.55843131 -0.7034877503
#> 57  057  1.42003509  0.1590656484
#> 58  058 -1.60903674 -1.7024415985
#> 59  059 -0.97370686 -0.8728941281
#> 60  060  0.64220693  1.3831492394
#> 61  061 -0.93962346 -1.8419127049
#> 62  062 -0.41546700 -1.2266189662
#> 63  063  1.31979990 -0.2639110819
#> 64  064  0.96041741 -0.3540629604
#> 65  065 -0.45483396  0.4225748580
#> 66  066  0.09394846 -0.2980900885
#> 67  067 -0.02129576 -0.2491092863
#> 68  068 -1.20881599  0.6734690403
#> 69  069 -0.58371931 -0.8612122966
#> 70  070  0.34454929  1.6862715094
#> 71  071  1.29629292  1.0699784160
#> 72  072  2.22339470  0.4363570733
#> 73  073 -1.12573953 -1.3565276587
#> 74  074  1.11665568  0.4968683355
#> 75  075  0.35950424  0.4817999803
#> 76  076  0.88910474 -0.2869615103
#> 77  077 -0.80015124  0.6908267447
#> 78  078 -1.13318986 -0.0132336550
#> 79  079 -1.37490376  0.1418901839
#> 80  080  1.26132278  0.7907948582
#> 81  081  0.08226323 -0.1451455402
#> 82  082  1.58878569  1.5698546905
#> 83  083  2.41458667  3.9993163493
#> 84  084 -0.79350230 -0.8956149616
#> 85  085 -0.28557458  0.8803869986
#> 86  086  0.11254159  1.2935555037
#> 87  087  1.95808148  0.8532413148
#> 88  088  0.96109821  2.3314268314
#> 89  089 -0.74813691 -0.4919982062
#> 90  090 -0.56784285  0.6716218704
#> 91  091 -1.07964448 -0.4387599896
#> 92  092 -0.11001263  0.8939155927
#> 93  093 -0.05618537  0.0818587309
#> 94  094 -0.91244017 -0.2167446451
#> 95  095 -0.23534401  0.4603693679
#> 96  096 -0.02752659 -0.6236095729
#> 97  097 -0.34331075  0.0433417234
#> 98  098  0.22193726 -0.6425720464
#> 99  099 -0.02787989 -0.8024229011
#> 100 100 -0.28243638 -1.4120794881