Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5179582
#> Y_2 0.5179582 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.2105857 -0.8531502
#> 2 -0.4840701  0.4321743
#> 3 -0.5166913 -0.9918149
#> 4  0.6370676  0.4164349
#> 5  0.4598923  0.2831948

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1           Y_2
#> 1   001  0.771280778  1.2049652472
#> 2   002  1.100676072  1.5726914612
#> 3   003 -0.983642746 -0.7270055700
#> 4   004  0.591137184 -0.0302460102
#> 5   005 -2.415176249 -0.2028158261
#> 6   006  1.945295346  1.4403813556
#> 7   007 -1.163300226 -0.2569318573
#> 8   008 -0.906761217 -0.4805911531
#> 9   009  0.404646647  0.7313905910
#> 10  010 -0.464769467 -1.3994990964
#> 11  011 -1.699370810  0.3969438300
#> 12  012  1.449220291  1.3145133289
#> 13  013  0.401775992 -0.5784885839
#> 14  014  1.350231648  1.6579793438
#> 15  015 -0.294851807 -0.0406869709
#> 16  016  1.333554103  0.7380388168
#> 17  017  1.911941578  0.4893101318
#> 18  018  0.345361020  0.7215733895
#> 19  019 -0.410029791  0.1699230913
#> 20  020  0.066773827 -0.9392418624
#> 21  021 -0.791664513  0.0228618441
#> 22  022 -0.089647138  0.4740789110
#> 23  023 -0.164502290  0.9173136733
#> 24  024  1.557782464  0.8600318936
#> 25  025  0.983739345  0.6400115264
#> 26  026 -0.257999486  0.5501433144
#> 27  027 -1.104106982 -1.1766041802
#> 28  028  1.899168721  0.6959825252
#> 29  029  0.034490531  0.7032098940
#> 30  030 -0.013968728 -0.0342075721
#> 31  031  0.619534896  1.2322992474
#> 32  032  0.622263279  1.2840124583
#> 33  033 -1.705183391 -0.5709846713
#> 34  034 -0.369067370 -0.3981902594
#> 35  035  1.257044515  0.5131654873
#> 36  036 -0.091634389  0.1399362208
#> 37  037  0.674981884  0.4939298363
#> 38  038  0.658752209  1.3210996280
#> 39  039  1.317133189  0.7222272937
#> 40  040  1.467171253  0.1596051756
#> 41  041 -1.491787373 -0.7265667150
#> 42  042  0.843922355 -1.0383217638
#> 43  043 -1.605959900 -0.2889236369
#> 44  044  1.531873592  0.5741219514
#> 45  045  1.760864061 -0.3652685265
#> 46  046  0.563854144 -0.2913144910
#> 47  047 -0.103057404  0.0006715828
#> 48  048 -1.810380116 -1.8194985322
#> 49  049  0.843330307  0.0939231243
#> 50  050  0.448903113  0.1966472471
#> 51  051  0.047367885  0.3985703463
#> 52  052  0.646790540  0.8974581839
#> 53  053  1.627942975  0.1181625253
#> 54  054 -1.464728889 -1.0820038224
#> 55  055  1.295483583  0.6221733163
#> 56  056 -0.093282847 -1.4500096921
#> 57  057  0.335398367  0.2659694781
#> 58  058  0.307623739  1.0501920539
#> 59  059  0.790335840  1.0435753389
#> 60  060  0.844979046  1.5726096672
#> 61  061 -1.957585473 -0.3181326418
#> 62  062 -0.076395007  0.9612026599
#> 63  063  0.253486192 -0.2255942323
#> 64  064 -0.743108315  0.0087108931
#> 65  065 -0.350047213 -0.0536020420
#> 66  066  0.584750848  0.1470824716
#> 67  067  1.639388032 -0.1691752770
#> 68  068 -0.109621948  0.3002299439
#> 69  069 -1.021308963 -1.0971799301
#> 70  070  2.245226136  2.7009297586
#> 71  071 -0.033434826  0.7438567945
#> 72  072 -0.845434176 -1.2796530485
#> 73  073 -0.398296893 -0.0716757690
#> 74  074  0.010224986  0.5986485907
#> 75  075 -1.124034368 -1.8718130329
#> 76  076  1.486624827  1.1759170286
#> 77  077 -0.917877852 -0.5562372153
#> 78  078 -0.423325266  0.8209686585
#> 79  079  0.333227995  0.1079764599
#> 80  080  1.735946083  1.4505385461
#> 81  081 -1.522102061 -1.2381895595
#> 82  082 -0.120268495 -2.4577407618
#> 83  083 -0.458898990 -1.7353226005
#> 84  084  1.777028997  1.0276680797
#> 85  085  0.371790511  0.1391208958
#> 86  086  0.296291352  0.1218094450
#> 87  087 -0.002908339  0.1220879772
#> 88  088  1.053719461 -0.5162970787
#> 89  089 -0.076584613 -1.2252085104
#> 90  090  0.089335487  0.5196732793
#> 91  091  0.951046957  0.0475582567
#> 92  092 -1.445620055 -1.1434160787
#> 93  093  1.501125999  0.9545860900
#> 94  094  0.187808087  0.0964038280
#> 95  095 -0.356631049 -2.0179124427
#> 96  096  0.489762346 -0.8482767764
#> 97  097 -0.399489130 -0.4891781740
#> 98  098 -0.227405994  0.4292854008
#> 99  099  0.706737397 -0.5417048865
#> 100 100  0.080089708  0.1422420543

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -1.15253639 -0.944126871
#> 2   002  1.22000798 -0.392523267
#> 3   003  2.24326394  2.939768542
#> 4   004 -0.08336737 -0.096504511
#> 5   005  0.46859505  0.336775185
#> 6   006  2.19888957  1.315731350
#> 7   007 -0.95384635 -0.475325427
#> 8   008 -0.63718123  0.078785856
#> 9   009  0.86131649 -1.683925645
#> 10  010  1.25388068 -0.890759391
#> 11  011  0.46810216  0.626135990
#> 12  012 -0.64792378 -1.178488805
#> 13  013  0.46521607 -0.422626123
#> 14  014  0.41163967 -0.193651838
#> 15  015 -1.02576340  0.991509985
#> 16  016 -1.26670736 -1.735579403
#> 17  017  0.35159889  0.114078352
#> 18  018  0.30603283  0.502093868
#> 19  019  1.47181892  0.793574245
#> 20  020  1.40890039  0.999014911
#> 21  021 -0.24459482 -0.460004503
#> 22  022  0.57066452 -0.769046936
#> 23  023  0.66316489 -0.344346969
#> 24  024 -1.66419480 -1.107674055
#> 25  025 -0.16169359  1.030384594
#> 26  026 -0.62853118 -0.222338658
#> 27  027 -0.59473326  0.024323511
#> 28  028 -0.14876793 -0.221781207
#> 29  029  0.34361785  0.430474021
#> 30  030 -0.54816240  1.126120609
#> 31  031 -1.78323291 -1.228611015
#> 32  032  0.20512899  0.076947845
#> 33  033 -0.30614309 -0.447006526
#> 34  034 -0.05539830  0.130352421
#> 35  035  0.09282943 -1.715575955
#> 36  036  1.34374086 -0.058412955
#> 37  037 -0.35815599 -0.276066187
#> 38  038 -0.54000199 -0.922581027
#> 39  039 -0.27812661  0.200357165
#> 40  040  0.02778713 -0.238870708
#> 41  041  1.18056212 -0.239549351
#> 42  042  0.03810477  0.936204631
#> 43  043 -0.07910439  0.475642586
#> 44  044 -0.71927129 -0.767827249
#> 45  045 -2.35765173 -2.041670016
#> 46  046 -0.05393377  1.514797828
#> 47  047  0.03620646  0.719610005
#> 48  048 -1.08584015 -1.584627110
#> 49  049 -0.22473367  0.366681176
#> 50  050 -1.98108452 -1.116161567
#> 51  051  0.25298945  1.867866705
#> 52  052  1.33949548  0.485883564
#> 53  053 -0.28600967 -0.202840490
#> 54  054  0.19496213 -0.125883973
#> 55  055 -0.24425590 -0.258877572
#> 56  056 -0.31162108  0.215671398
#> 57  057 -1.35544174 -0.676588082
#> 58  058  0.79536349  0.952381098
#> 59  059  1.85046815  1.352999320
#> 60  060 -0.96320645 -1.078265359
#> 61  061  0.59940337  1.929157795
#> 62  062  0.27262469 -0.948034706
#> 63  063 -0.60356925 -0.555453736
#> 64  064 -0.02056769 -0.319220958
#> 65  065 -0.09819895  0.757855846
#> 66  066  2.34900583  1.939061004
#> 67  067 -0.01659304 -0.715734483
#> 68  068  0.41926980  1.520735103
#> 69  069 -1.73353601 -1.458736295
#> 70  070 -0.60353890 -0.628945969
#> 71  071  0.31322253 -1.001737281
#> 72  072 -0.61030881 -0.188330638
#> 73  073 -1.62182091 -0.926008170
#> 74  074 -0.93071359 -1.107430047
#> 75  075  1.54385035  1.115070702
#> 76  076 -0.64190136 -0.248995124
#> 77  077 -0.63962206 -0.470274798
#> 78  078  0.49488746 -0.008829213
#> 79  079 -0.91715243 -0.088078932
#> 80  080  0.34316395  0.413084024
#> 81  081 -1.40217113 -0.521669349
#> 82  082  0.94127032  0.716075579
#> 83  083 -0.25477645  0.708391731
#> 84  084  0.47359365  2.703917674
#> 85  085 -1.23733662 -0.244750838
#> 86  086 -1.15344888 -0.156218254
#> 87  087  0.02987274 -1.338560592
#> 88  088  1.28502328 -0.047692216
#> 89  089 -1.46362678  0.840973509
#> 90  090  1.25258428  0.495036937
#> 91  091 -1.40024683 -2.336600283
#> 92  092  0.21222681  0.544368507
#> 93  093 -0.06556563  0.795184084
#> 94  094 -0.55405587  0.590740416
#> 95  095  0.24682535 -0.231832831
#> 96  096  2.09096608  1.101436850
#> 97  097 -1.53554832 -0.967233551
#> 98  098  0.42269931  0.400458664
#> 99  099  1.01503702  2.061197960
#> 100 100 -1.71881113 -0.133522518