Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4995189
#> Y_2 0.4995189 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1940711 1.06931437
#> 2 -0.6604037 0.58378584
#> 3 -1.2795472 0.07887152
#> 4  1.1725985 0.46533902
#> 5  0.4144625 0.12766771

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  1.07149608  2.164469747
#> 2   002 -0.61822633 -0.679550637
#> 3   003  0.07573049 -0.445603127
#> 4   004  0.40142025  0.452592134
#> 5   005 -0.91677723 -0.401773785
#> 6   006 -1.20176852 -0.491020784
#> 7   007  1.37730587  0.708464724
#> 8   008 -0.43835837  0.452435647
#> 9   009  0.63095746 -0.007057666
#> 10  010  0.59946508  0.913596203
#> 11  011 -0.52134851  0.703781002
#> 12  012  0.90341920  0.226642687
#> 13  013  2.06795530  1.164094489
#> 14  014  1.37367660 -0.283395546
#> 15  015  1.63980622 -0.028329548
#> 16  016 -0.56443734 -0.067588467
#> 17  017 -0.81226285 -1.238654780
#> 18  018  0.01440416  0.534344846
#> 19  019 -0.33812503 -0.460326535
#> 20  020 -0.71210677 -1.644935313
#> 21  021 -0.91092056 -0.137861406
#> 22  022 -1.12584190 -0.661848553
#> 23  023  0.70050046  0.571757671
#> 24  024  0.65394560  1.082698132
#> 25  025  0.79619048  0.636319793
#> 26  026  0.36916165 -0.564391418
#> 27  027  1.39924146  0.189311927
#> 28  028  0.11947097 -0.278359877
#> 29  029  0.38813163  0.504218551
#> 30  030 -0.18681387 -0.215339117
#> 31  031 -1.61482934 -0.771914188
#> 32  032  1.12701356 -0.316023658
#> 33  033 -0.90593003  0.142192896
#> 34  034 -1.00545247 -1.297309144
#> 35  035  1.65073682  0.738291597
#> 36  036 -1.68812138 -0.612242236
#> 37  037  1.44615363 -0.518431092
#> 38  038 -0.15992966 -1.308591595
#> 39  039  1.30476154 -0.670248800
#> 40  040 -0.37059489 -0.451297551
#> 41  041  0.98473736  1.274041961
#> 42  042 -0.61227579 -0.458046979
#> 43  043 -0.25036431 -0.775511272
#> 44  044 -0.37490998 -1.853071764
#> 45  045  0.67935592  0.322865266
#> 46  046 -1.05408702  0.730903299
#> 47  047  0.20333552 -0.639447799
#> 48  048 -0.78645812 -1.146707457
#> 49  049 -1.80617082 -0.401334054
#> 50  050 -0.75883310 -1.737778422
#> 51  051  1.60476532  0.607350440
#> 52  052  0.76746347 -0.374581098
#> 53  053  0.15903289  0.096709084
#> 54  054 -0.32647455 -0.290041794
#> 55  055  0.23260345 -0.246470447
#> 56  056  0.60407886  0.324758337
#> 57  057  1.13521461  0.320356431
#> 58  058  1.10747373 -0.618524931
#> 59  059 -1.66644138 -0.452944414
#> 60  060  1.21747949  2.379462285
#> 61  061  0.18126436 -1.094957328
#> 62  062 -0.57687140 -0.496575697
#> 63  063  0.56602339 -0.015232368
#> 64  064  0.58472211 -1.219793333
#> 65  065 -1.03099003  0.096181816
#> 66  066  0.44995603 -1.039157702
#> 67  067  1.23833130  2.224351687
#> 68  068  0.63380948 -1.346094703
#> 69  069  0.27961062 -0.627604366
#> 70  070 -1.64789516 -0.245423116
#> 71  071  1.01762769  1.236287196
#> 72  072 -0.61533988  0.349306578
#> 73  073 -0.32070042 -1.304770695
#> 74  074 -1.23546201 -1.052911729
#> 75  075  0.84007381  1.284834978
#> 76  076  0.24242426  0.722068011
#> 77  077  1.46774158  0.247560981
#> 78  078  0.62421068  0.533950061
#> 79  079 -1.03073049 -0.524443354
#> 80  080  0.92715367  0.453223851
#> 81  081  0.37127802 -0.414608367
#> 82  082 -0.40217522 -0.887785303
#> 83  083 -0.93127216  0.344009474
#> 84  084  0.14212917  0.752770384
#> 85  085 -0.47170262 -0.573836769
#> 86  086 -0.57514828 -0.582580513
#> 87  087 -0.93988654 -0.591083448
#> 88  088 -0.72465880 -1.065845277
#> 89  089 -0.89839649 -0.477352569
#> 90  090  1.04441723  0.789840120
#> 91  091 -0.18714724 -1.365692436
#> 92  092  1.07022585  1.675566797
#> 93  093 -0.65935250  0.636134977
#> 94  094 -0.18663281 -0.009905018
#> 95  095 -0.68784530  0.086173667
#> 96  096  1.43813846 -0.093516138
#> 97  097 -1.36397852  0.778845981
#> 98  098 -1.44982390  0.120579856
#> 99  099  0.11817521 -1.247693343
#> 100 100  0.52416338  1.350097603

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001  0.060644344  0.28900634
#> 2   002 -0.557720466  1.14010097
#> 3   003  0.227778445  1.04562058
#> 4   004  0.676232101  0.71770069
#> 5   005  0.071505786 -1.42445606
#> 6   006 -0.583080641 -0.34000673
#> 7   007 -0.872689030 -0.03187165
#> 8   008 -1.154148676  0.55494985
#> 9   009  2.591671363  1.51505960
#> 10  010  1.104872518 -0.31584586
#> 11  011  1.332495854  0.62800607
#> 12  012 -0.891618310 -0.11898506
#> 13  013 -0.265182116 -0.07600564
#> 14  014  0.503287478  0.73857110
#> 15  015 -1.097846842  0.45311411
#> 16  016  0.014035445 -0.26414155
#> 17  017 -0.474576260  1.63939908
#> 18  018 -0.223597324  0.07730556
#> 19  019 -0.184819938 -0.19557123
#> 20  020  0.678250922  1.54713382
#> 21  021  1.462394488  0.34321388
#> 22  022 -1.609058261 -1.34075166
#> 23  023 -0.485329446  0.85532996
#> 24  024  0.241633491 -2.47525430
#> 25  025 -1.586507657  0.21356531
#> 26  026 -1.789795545  0.38002800
#> 27  027  0.388610215  0.91390744
#> 28  028 -1.508746592 -0.98076068
#> 29  029  1.323333131  0.70603512
#> 30  030 -0.852096661 -0.33893478
#> 31  031 -1.640517288 -0.13156954
#> 32  032  0.597886715 -0.70885148
#> 33  033 -1.106090909  0.25869057
#> 34  034  1.136068849  1.74280642
#> 35  035  0.185093603 -0.07467841
#> 36  036  0.399154898 -0.01990701
#> 37  037 -0.462005050  0.18655319
#> 38  038 -0.415629516  0.69117230
#> 39  039 -1.411976055 -0.30982943
#> 40  040  0.364750783  0.07034510
#> 41  041 -0.155322343 -1.75746483
#> 42  042 -0.053678725  0.47263797
#> 43  043 -0.321956255  0.74583532
#> 44  044 -1.146690571 -0.89607084
#> 45  045  1.363414710  1.74675015
#> 46  046 -0.406685194 -0.14358160
#> 47  047  1.216000979  0.04936456
#> 48  048 -0.144091506 -0.66599039
#> 49  049  0.833492379 -1.63380317
#> 50  050 -0.281413970 -0.65562966
#> 51  051  1.952171265  1.15805837
#> 52  052  0.070261508  0.93513825
#> 53  053 -0.474580038 -1.04393959
#> 54  054 -0.421969657 -0.14915139
#> 55  055 -0.497260942  0.03271817
#> 56  056 -0.587002477 -1.72838659
#> 57  057 -1.527392063 -0.39150878
#> 58  058 -0.203844249 -0.33735663
#> 59  059  0.736866193  0.84267282
#> 60  060 -0.040007759 -1.94284179
#> 61  061  0.476133580 -0.51044639
#> 62  062  0.997729614  0.56109573
#> 63  063  1.014942395 -0.20739237
#> 64  064 -1.846905660 -1.15422876
#> 65  065 -0.966241335  0.36036569
#> 66  066 -0.414509986 -0.63184022
#> 67  067  0.264463278  1.23677477
#> 68  068 -0.617781767 -1.85362352
#> 69  069  0.696813789 -1.38880146
#> 70  070 -0.078756321 -0.89612643
#> 71  071  0.758157528  0.69454289
#> 72  072 -0.745108143 -1.26867817
#> 73  073  0.916499464  0.91426145
#> 74  074 -0.167623679  0.08375816
#> 75  075 -0.745687168 -1.08760775
#> 76  076 -1.234576995 -2.02870518
#> 77  077  1.481435734  0.11171256
#> 78  078  0.795758970 -0.52138223
#> 79  079 -0.363773563  0.90670154
#> 80  080 -2.672391086 -1.21426812
#> 81  081  0.194513386  0.80400027
#> 82  082  1.254864595 -0.25413482
#> 83  083  0.285187356  0.04957569
#> 84  084 -0.770746744 -0.34793555
#> 85  085  1.632544805  0.99015485
#> 86  086  0.063066739 -0.48347477
#> 87  087  0.325237292  1.36476794
#> 88  088 -1.014815830  0.85677255
#> 89  089  0.037869137  0.08017518
#> 90  090 -0.263135745 -0.29439977
#> 91  091 -0.506020639 -1.60235355
#> 92  092 -0.361334841 -0.62710494
#> 93  093  0.074377733 -0.72835777
#> 94  094  0.775866161 -0.01858959
#> 95  095  0.166709773  1.04085707
#> 96  096 -0.094974962 -0.86698400
#> 97  097  1.130451631  1.15764962
#> 98  098  1.888674679  1.23873377
#> 99  099 -0.009756032 -0.56444346
#> 100 100  0.273589594 -0.63337882