Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4979209
#> Y_2 0.4979209 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  2.1959585  0.2751867
#> 2  0.6612760  1.4778320
#> 3 -0.4270942  0.5456716
#> 4  1.0369592  0.6317161
#> 5 -1.0380917 -1.3214657

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -2.311387238 -1.715549037
#> 2   002  0.298080446 -0.882494346
#> 3   003 -1.310049678 -0.003932881
#> 4   004 -0.769531255 -0.705692717
#> 5   005  1.175759641  2.083647510
#> 6   006 -0.245537689  1.047876641
#> 7   007 -0.337073659  0.282990381
#> 8   008  2.731519282  1.785833612
#> 9   009  1.273771499  0.535919730
#> 10  010  1.528288972  0.784902617
#> 11  011  2.098289500  1.373094659
#> 12  012  0.076985978 -0.223146069
#> 13  013  0.844993651 -0.346253428
#> 14  014  0.399626166 -0.462898232
#> 15  015  0.164879826  1.139990618
#> 16  016 -1.326918562 -1.826008843
#> 17  017 -0.138194313  0.254049164
#> 18  018  0.747214134  0.992208390
#> 19  019  0.638601339 -0.408391222
#> 20  020 -0.808167448 -0.921161514
#> 21  021 -0.604279273  0.007892215
#> 22  022 -0.506637349 -1.018173922
#> 23  023  0.508214547 -0.738873698
#> 24  024 -0.989405418 -1.054174704
#> 25  025  0.051898427  0.478031774
#> 26  026  0.389802961  0.319370678
#> 27  027 -1.428426136 -0.634122287
#> 28  028  0.023507195 -0.464613207
#> 29  029  0.058345063  1.299650222
#> 30  030 -1.309317600 -1.925551652
#> 31  031  1.546246853 -0.305029396
#> 32  032 -1.006229721 -1.158304186
#> 33  033 -0.140783974  1.219327584
#> 34  034 -1.783108851 -1.550030336
#> 35  035  0.390444695  1.821992995
#> 36  036  0.026650039 -1.758706795
#> 37  037  0.012030346  0.028547668
#> 38  038 -1.361987450 -0.804554200
#> 39  039  1.510613657  0.955801731
#> 40  040  0.312658273 -1.438145226
#> 41  041 -1.084311534 -0.216314761
#> 42  042  0.721628850 -0.082360076
#> 43  043 -0.160415839  0.094258427
#> 44  044 -0.745271987  0.647471272
#> 45  045  0.811014166  1.209653531
#> 46  046 -0.341253280  0.169092007
#> 47  047  0.617731743  0.616634264
#> 48  048  0.061413430  0.309462145
#> 49  049  0.641206790  0.220128891
#> 50  050  0.736590484  0.465980249
#> 51  051  0.820749082 -0.087549615
#> 52  052  1.283205860 -0.225983057
#> 53  053  0.198670735  1.052284038
#> 54  054 -0.003603666 -0.836047430
#> 55  055  0.111293434  0.847090788
#> 56  056  0.134665479 -1.810837027
#> 57  057  0.678205016  0.235345731
#> 58  058  0.571045498  0.248040717
#> 59  059  4.058495386  2.368080306
#> 60  060  2.307618394  0.950936020
#> 61  061 -0.246155053  0.593464959
#> 62  062 -0.749038664 -0.418068982
#> 63  063  0.251838108 -0.585930524
#> 64  064  1.260018880  0.791618689
#> 65  065  0.603508739  0.099909393
#> 66  066  1.502038471  1.684363303
#> 67  067 -0.057609104  1.390681683
#> 68  068 -1.020929729 -1.184890318
#> 69  069 -0.717663577 -1.520930038
#> 70  070 -0.754615937 -1.175130412
#> 71  071 -1.202948756 -0.874102662
#> 72  072  1.595969273 -0.301065340
#> 73  073 -1.098823752 -2.089959879
#> 74  074  1.993812665  1.770685360
#> 75  075 -1.426131917 -2.056035281
#> 76  076  0.276027240  0.677351985
#> 77  077 -0.085043220 -0.693029283
#> 78  078 -0.530084420  0.506375938
#> 79  079  1.654636951  1.369033221
#> 80  080  1.815861208  0.372307904
#> 81  081 -0.039020491 -0.903566625
#> 82  082  0.390987194  0.082827121
#> 83  083  1.104275101  0.109457861
#> 84  084 -0.949025616 -1.298990778
#> 85  085 -0.986536230 -0.580804829
#> 86  086  0.845647702 -0.125289218
#> 87  087 -0.307623759 -1.260768555
#> 88  088 -1.265692848 -0.095442717
#> 89  089 -0.271040932 -0.982304715
#> 90  090  0.093451637  0.367430653
#> 91  091  0.895469594  1.063052364
#> 92  092  0.051286473  0.927592022
#> 93  093  0.451086655 -1.288878256
#> 94  094  2.035494409  2.889770967
#> 95  095 -0.556392789 -0.240885909
#> 96  096 -1.851823447 -0.797862645
#> 97  097  1.505265162  0.742723257
#> 98  098  0.498968706  0.436148843
#> 99  099 -0.853030858 -1.247175762
#> 100 100  0.350391696 -0.206464148

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.53815308  0.237376869
#> 2   002  1.33212454 -0.242967681
#> 3   003 -0.62650867 -0.126509679
#> 4   004 -0.26200653  0.507109330
#> 5   005  0.65210909  0.270780723
#> 6   006 -0.21456283 -0.582421960
#> 7   007  0.26850937 -0.856143894
#> 8   008  0.96176330  0.393733680
#> 9   009 -0.32481462 -0.123784747
#> 10  010  1.52852378  0.372211076
#> 11  011  0.21360253 -0.502966755
#> 12  012  1.42867604  0.868966729
#> 13  013  1.10173748 -0.064617069
#> 14  014  3.49892542  2.658526023
#> 15  015 -0.16857806 -0.066617023
#> 16  016 -1.24861458  0.899907638
#> 17  017  0.60831204  0.907829803
#> 18  018 -0.81469456  0.701618000
#> 19  019  0.63616679  0.879766156
#> 20  020 -0.01996582  0.608706692
#> 21  021  0.31559611  0.849795406
#> 22  022  1.02616753  1.248947071
#> 23  023  1.21273165  1.814232562
#> 24  024 -0.48456560  0.357389876
#> 25  025  1.06081278  2.506189988
#> 26  026  0.45885528  1.669660370
#> 27  027 -1.30014033 -0.710955671
#> 28  028  0.92882557  1.120276330
#> 29  029 -0.38853964 -0.058775119
#> 30  030  2.06311282  1.129918501
#> 31  031  0.60065028  0.363867925
#> 32  032 -2.15028758 -0.008739074
#> 33  033  0.72268134  0.228365400
#> 34  034 -1.24272609  0.747004066
#> 35  035 -0.29499424 -0.026518214
#> 36  036 -0.48874248 -0.619439572
#> 37  037  0.74225744  1.881845708
#> 38  038 -0.18426884 -0.684709529
#> 39  039  1.01783496  0.242016158
#> 40  040 -0.79005771  1.371261370
#> 41  041  0.07961180 -1.282271535
#> 42  042  0.05973796 -1.588539587
#> 43  043  0.14073706 -1.136725361
#> 44  044  1.47102660  0.966172438
#> 45  045  1.05853464 -0.210770809
#> 46  046  0.70369343  0.089219647
#> 47  047 -0.73429211 -0.246460645
#> 48  048  2.27626377  1.883439369
#> 49  049 -1.96710364  0.015407526
#> 50  050 -0.52686269 -0.904561247
#> 51  051  0.03154320 -0.733343878
#> 52  052  0.81013194 -1.041008706
#> 53  053 -0.05327538 -0.914153606
#> 54  054  0.26711708  1.034046644
#> 55  055  0.38497978  0.076692475
#> 56  056 -1.55229750 -1.680265004
#> 57  057 -2.79359235 -1.257058780
#> 58  058 -0.29270730  0.302515123
#> 59  059  1.19876112  0.601320067
#> 60  060 -0.42853625 -0.570453019
#> 61  061 -0.15675232  0.654361756
#> 62  062  0.50199740  0.107864015
#> 63  063  0.66560160  0.711723763
#> 64  064  0.75321728  0.049049397
#> 65  065  0.33987446  1.388170950
#> 66  066  0.07920970  0.091138429
#> 67  067  1.40367538  0.683065594
#> 68  068  2.21239692  0.965995849
#> 69  069 -0.11794665  0.515330512
#> 70  070  0.31996034 -0.225802765
#> 71  071  1.67056561 -0.486733476
#> 72  072 -0.50902748 -0.217821576
#> 73  073 -0.58749820 -0.150492830
#> 74  074 -0.33498490 -0.040515163
#> 75  075  0.31674030  0.386803777
#> 76  076  1.00141744  0.644021185
#> 77  077  0.16432632  0.066924877
#> 78  078 -1.11116920 -2.017331631
#> 79  079  0.95052845  1.852395528
#> 80  080 -0.79312434 -0.529368456
#> 81  081  1.19167508  0.483663977
#> 82  082 -0.32847956 -0.631031322
#> 83  083 -0.06201703  0.062431260
#> 84  084 -1.26865359 -0.063795265
#> 85  085  0.65711823  0.265088268
#> 86  086 -1.06044690 -0.229458537
#> 87  087  1.15986533  1.165676629
#> 88  088 -0.89131120  0.158874568
#> 89  089 -0.81006403  0.740162462
#> 90  090  0.80731379  1.663125654
#> 91  091  2.29191982  2.216504248
#> 92  092 -0.59787398 -0.297321473
#> 93  093 -0.76239590 -0.333334182
#> 94  094 -1.20419012 -0.534110604
#> 95  095 -0.22017870  1.691346547
#> 96  096  3.35785390  2.173344012
#> 97  097  0.90267865 -1.626438979
#> 98  098  0.33311962 -1.127851703
#> 99  099  0.38341865  1.111149300
#> 100 100  1.23874907  1.123756614