Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5312106
#> Y_2 0.5312106 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1       Y_2
#> 1  0.68014862 1.4274393
#> 2  0.03189694 0.1050024
#> 3 -0.18221734 0.1162269
#> 4  1.40211648 1.2831821
#> 5  0.18966395 1.4206567

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.10406751  0.34154848
#> 2   002  0.29658527  0.46428717
#> 3   003 -0.10731878  0.88711281
#> 4   004 -0.30087176  0.40729820
#> 5   005 -0.69709472 -0.30714562
#> 6   006 -0.10278353 -1.20748935
#> 7   007  0.73395238  0.81441691
#> 8   008 -2.39476807  0.92155969
#> 9   009 -1.53201261 -1.32663985
#> 10  010 -0.03084980 -0.25902390
#> 11  011 -0.67707735 -0.54077331
#> 12  012  2.36949802  1.50347875
#> 13  013 -1.42831726 -1.19350188
#> 14  014  0.23663981  0.86470066
#> 15  015 -0.23981502  0.37018500
#> 16  016 -0.37921249 -0.13158472
#> 17  017  0.21069754  0.21286458
#> 18  018 -0.52323394 -2.04944265
#> 19  019  1.27529274  1.71197754
#> 20  020  0.30853242 -0.61359687
#> 21  021 -0.83239794  0.18897288
#> 22  022  0.04047195  1.13948351
#> 23  023  0.95009420  2.26184609
#> 24  024 -0.51630291 -0.28154687
#> 25  025 -0.74857451 -0.71975531
#> 26  026 -0.17439498  0.26516751
#> 27  027 -1.20258049  0.82478312
#> 28  028  0.79028158 -0.32986322
#> 29  029 -1.25249202  0.95774557
#> 30  030  0.04878270 -2.40200918
#> 31  031  2.69486464  2.85907682
#> 32  032  0.28695131 -0.57969973
#> 33  033  0.70097454 -0.01216604
#> 34  034  0.20944850  0.97032827
#> 35  035  0.40692208 -0.37319499
#> 36  036  1.55816433  1.23435275
#> 37  037  0.20203530  0.33422395
#> 38  038 -0.35506648  0.55218248
#> 39  039  0.27475702 -0.25017425
#> 40  040 -0.25611849  0.62339955
#> 41  041  0.28233122 -0.62513418
#> 42  042 -0.71242962  1.45430206
#> 43  043  1.09752073 -0.38012132
#> 44  044 -0.74967096  0.55529004
#> 45  045 -0.02929288  0.65626052
#> 46  046  0.39271603 -0.69164721
#> 47  047 -1.45712201 -2.12038894
#> 48  048  1.81482841  0.18462339
#> 49  049 -0.61433249  0.24775378
#> 50  050 -1.61953970 -0.78691947
#> 51  051 -0.19103955 -0.80466962
#> 52  052 -2.38586653 -1.62042474
#> 53  053  0.73200273  1.30328772
#> 54  054 -0.30711866  0.48736618
#> 55  055 -0.82873025  1.50501499
#> 56  056  0.37828064 -0.10647539
#> 57  057 -1.44846874 -1.61951944
#> 58  058  0.45327718 -0.49785759
#> 59  059  1.13924935 -0.29539051
#> 60  060  1.17818907  0.62799201
#> 61  061  0.10077025 -0.23588286
#> 62  062 -1.08696428  0.23579051
#> 63  063  1.27595490 -0.20799685
#> 64  064 -0.40771690  0.16325210
#> 65  065 -0.45361014  0.07808402
#> 66  066  0.69401380  0.97455118
#> 67  067  0.32355003  0.60459488
#> 68  068  0.17087678 -2.60153980
#> 69  069 -0.39449759  0.88734248
#> 70  070  0.49814750 -0.57879285
#> 71  071  0.82575857  0.81689476
#> 72  072  1.31185839  0.71233045
#> 73  073 -0.31004218  0.06221998
#> 74  074  0.00584620 -1.47054161
#> 75  075 -1.88532155 -2.57733930
#> 76  076  1.27990011  0.84611776
#> 77  077 -0.79222490 -0.39794551
#> 78  078  0.70251874  1.04376080
#> 79  079 -0.71942663 -0.60574478
#> 80  080  0.23015853 -0.97619438
#> 81  081 -1.03053491 -1.41467716
#> 82  082  0.04164499  0.61269660
#> 83  083 -1.51291363 -1.17040151
#> 84  084 -1.13936000  0.63066524
#> 85  085 -0.81214538 -0.88098529
#> 86  086 -2.02201748 -0.55992682
#> 87  087  0.08679007 -0.97170291
#> 88  088 -0.17973083 -0.60111308
#> 89  089  1.86440128  0.19455503
#> 90  090 -0.41053183  0.50534956
#> 91  091 -1.98897199 -2.56341422
#> 92  092 -1.06915008 -1.64646450
#> 93  093 -1.27301922 -0.66422227
#> 94  094  1.00867873  0.47178204
#> 95  095  0.29400741  1.07327358
#> 96  096 -0.25299058  0.97176045
#> 97  097 -0.34585425 -0.33600950
#> 98  098  1.06414483  0.65712791
#> 99  099 -2.55659498 -2.72541324
#> 100 100 -0.35056306  0.52582683

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001  1.0583948381  1.24353737
#> 2   002 -0.6864304170  0.59689050
#> 3   003  1.0869485706  0.09883028
#> 4   004  2.1116179765  1.81292118
#> 5   005  0.8962619153  0.06598800
#> 6   006  0.1598774655 -1.18559868
#> 7   007  0.7906246281  1.11512139
#> 8   008 -0.6382070244  0.56037955
#> 9   009  1.1150208999  1.04179432
#> 10  010  1.1726841517 -0.10281657
#> 11  011  1.1776053906 -0.23143406
#> 12  012  1.5421165885  0.72556671
#> 13  013  0.1853038376 -0.09958824
#> 14  014  0.2448931525 -0.04017268
#> 15  015  0.3363599123  0.98893103
#> 16  016 -0.2327129977 -0.25486467
#> 17  017 -0.4900297952 -0.49521910
#> 18  018  0.7547098981  0.69309561
#> 19  019  0.9752077419  0.51851486
#> 20  020 -0.9538032256  0.96369945
#> 21  021 -1.2586671588 -0.29368365
#> 22  022  1.0294508264  0.36479416
#> 23  023 -0.0121605931  0.46524723
#> 24  024 -1.7649405531 -0.56894263
#> 25  025  0.6844594275 -0.81113205
#> 26  026  1.5650229239  1.40984184
#> 27  027 -0.0507956264  1.21608914
#> 28  028  1.0929445119  0.52241256
#> 29  029  0.1463782861  1.59331912
#> 30  030 -1.4796204154 -1.05320191
#> 31  031  1.0437257618  0.78701769
#> 32  032  0.0035528124  0.19576082
#> 33  033 -0.7012701593  0.45622306
#> 34  034  1.2011037699 -0.74388344
#> 35  035  0.5218874619  0.97051148
#> 36  036  1.1052386130  1.01936817
#> 37  037  0.1849519517  0.46937061
#> 38  038  0.7320289505 -0.56893757
#> 39  039  1.7035381557  1.35659498
#> 40  040 -0.9321985133  0.91294748
#> 41  041  0.3179435445  1.47154872
#> 42  042 -0.0219524904  0.16734879
#> 43  043  0.1719787462  0.38259737
#> 44  044  0.2917000428 -0.38665654
#> 45  045  0.9294143720  1.12785369
#> 46  046  0.9303770332 -0.08417072
#> 47  047  1.4100527014  1.34532480
#> 48  048 -0.3134694395 -1.43653989
#> 49  049  1.0133489545 -0.93372168
#> 50  050  0.6536875308 -0.11644713
#> 51  051  0.0909570383 -0.56170097
#> 52  052 -0.5317832565  0.56470360
#> 53  053  0.2539713352  0.85583097
#> 54  054  2.2253808988  0.38525228
#> 55  055 -0.5970758858  0.18411209
#> 56  056 -0.5048021301 -0.81678041
#> 57  057 -1.5847868764 -0.43390257
#> 58  058  1.3634825126  1.59839640
#> 59  059 -2.8230864325 -2.20109758
#> 60  060 -0.7868984771 -0.98026392
#> 61  061  0.6407502397  2.19470986
#> 62  062  0.0765130636  0.79972273
#> 63  063 -2.7133617971  0.21288562
#> 64  064 -1.2139001129 -0.43426050
#> 65  065 -1.7804339203 -1.05049276
#> 66  066 -1.5197187134 -1.61000900
#> 67  067 -0.4308594835 -0.20432299
#> 68  068  0.2853066383  0.50131312
#> 69  069  0.1413335488  0.71198816
#> 70  070  0.4292231752 -0.99410843
#> 71  071 -0.9635909956  0.16438908
#> 72  072 -1.3398693849 -1.66308571
#> 73  073  2.1708889419  0.65821743
#> 74  074 -0.6437622375 -0.07711923
#> 75  075 -0.0791828173 -1.15920576
#> 76  076  0.4304349661 -1.27884231
#> 77  077 -0.2259750974  0.76781191
#> 78  078 -0.4360090222 -0.92977524
#> 79  079  1.1338391300 -0.01854645
#> 80  080  1.0854696017  0.68592578
#> 81  081 -0.0008337915  0.35468663
#> 82  082 -0.4128660005  0.27337660
#> 83  083 -0.2187547230  0.34495144
#> 84  084  1.8694555403  1.77216095
#> 85  085  1.0239864999 -0.36817962
#> 86  086 -0.0462296221 -0.80465704
#> 87  087  0.3183814492  0.36509064
#> 88  088 -0.6747212134 -0.18698052
#> 89  089 -0.6275503587 -1.74197307
#> 90  090  0.0137849489 -0.89592526
#> 91  091 -0.5051998373 -0.45152537
#> 92  092  0.6592355275  0.49813695
#> 93  093  0.4175215069  0.17511000
#> 94  094 -0.2369440728  1.74585150
#> 95  095 -0.1809399177  1.06838221
#> 96  096  1.7981000602  0.77113455
#> 97  097 -0.0191558168 -0.81526301
#> 98  098  0.1196990239  1.00554839
#> 99  099 -0.2972828637 -0.34241690
#> 100 100  1.5193254650  1.20022951