Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4784838
#> Y_2 0.4784838 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1  1.0527951  1.09163583
#> 2 -0.3492431 -0.30560074
#> 3  0.4312963 -0.37366421
#> 4 -0.4635413 -1.08500439
#> 5 -0.3015462 -0.01850103

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.561998163 -0.519193299
#> 2   002  0.172408892 -0.793884322
#> 3   003 -0.006310475  2.076255398
#> 4   004  0.486015799  1.356672966
#> 5   005  0.268426219 -0.287947817
#> 6   006 -0.036754284 -0.847045745
#> 7   007  0.836770699  0.093026052
#> 8   008 -0.546487103 -0.392793669
#> 9   009  1.997996246  1.828583779
#> 10  010  0.143053954  1.764476981
#> 11  011 -0.058529289 -0.800991064
#> 12  012 -0.412828116  0.539620294
#> 13  013 -0.058144249  0.020409305
#> 14  014  0.437989237  0.592760086
#> 15  015 -1.081126069 -1.089425612
#> 16  016 -1.201782072  0.081703526
#> 17  017  0.488793500 -1.029223560
#> 18  018 -2.618424452 -1.433122152
#> 19  019  1.493320988  0.245037852
#> 20  020  1.151324287  0.703715739
#> 21  021 -1.784899498 -0.962558267
#> 22  022  1.074726904 -2.204033303
#> 23  023  0.138852277 -0.697826111
#> 24  024 -1.097906288 -0.424170193
#> 25  025 -0.113851175 -0.599550590
#> 26  026  1.595420055  2.450658763
#> 27  027  1.986232314  1.235898122
#> 28  028  1.775935841 -0.778950088
#> 29  029 -0.964023268 -0.905877547
#> 30  030  0.355355183  0.622067092
#> 31  031  1.172026353  1.108697082
#> 32  032  1.555309587  1.066875391
#> 33  033  1.116059615 -1.015635962
#> 34  034 -0.895147139 -0.961752417
#> 35  035 -0.504884047 -1.544530412
#> 36  036 -1.301950232 -2.032126303
#> 37  037 -0.240683388 -1.511311324
#> 38  038  0.395027883 -0.116718742
#> 39  039  0.642790145 -0.374878622
#> 40  040 -1.764796737 -0.064451115
#> 41  041 -1.983256326 -1.009596197
#> 42  042  0.626820569  0.155215031
#> 43  043 -0.008305445  0.023578210
#> 44  044  0.728736555  0.083878807
#> 45  045 -1.469072963 -1.823371973
#> 46  046  1.808083326  0.691209099
#> 47  047 -1.029070051  0.446125676
#> 48  048  0.944738189  0.139541985
#> 49  049  0.260753508 -0.142164206
#> 50  050  0.513405825  1.746857745
#> 51  051  0.896747657  1.799768575
#> 52  052 -0.412093835 -0.126475923
#> 53  053  0.919083870  0.205817421
#> 54  054  0.506662107  0.055535809
#> 55  055  0.658501615  1.356835945
#> 56  056 -1.787254455 -0.627351977
#> 57  057 -0.120902366 -0.696659764
#> 58  058 -1.655545989 -1.231280483
#> 59  059  1.033846457  1.075581364
#> 60  060  2.027902442  1.243318724
#> 61  061  0.557246869 -0.252937065
#> 62  062 -0.373473001  0.443649137
#> 63  063 -0.185221525  0.221369669
#> 64  064 -0.473177804  0.327328094
#> 65  065  1.068708927 -0.265651066
#> 66  066  1.113839741  1.480546066
#> 67  067 -0.853681665 -0.873796800
#> 68  068  0.065401056 -1.217027729
#> 69  069 -0.029602445  0.696676854
#> 70  070  0.269627568 -0.438933933
#> 71  071 -1.431935633 -1.413190615
#> 72  072 -0.099782412 -0.335631513
#> 73  073  1.152918834  0.922768965
#> 74  074 -0.286145047  0.068509649
#> 75  075  1.262792608  0.038475377
#> 76  076 -0.092590029 -0.950363143
#> 77  077  0.752257811  0.279515703
#> 78  078 -0.755598822 -1.885023695
#> 79  079 -0.985544731 -0.114119404
#> 80  080  2.024443370 -0.426633594
#> 81  081  0.390152211  0.403767810
#> 82  082 -1.117858681 -0.682309808
#> 83  083  1.635363551  0.688098759
#> 84  084 -0.245264372  0.346346716
#> 85  085 -0.502802114  0.094017526
#> 86  086  1.672711845  0.830981838
#> 87  087  0.637185314 -0.306291026
#> 88  088  2.152105943  1.147369793
#> 89  089  0.824448339  1.481257118
#> 90  090  0.443540437  0.318346010
#> 91  091  2.600994138  0.237970573
#> 92  092  1.070151534 -0.333964573
#> 93  093 -0.594221600 -0.410633766
#> 94  094 -0.893144341  0.008673417
#> 95  095 -1.155063674  0.252425600
#> 96  096  0.899373350  0.637921142
#> 97  097  0.075819084 -0.966346305
#> 98  098 -1.181696374 -1.002195733
#> 99  099 -0.806211788 -0.252210924
#> 100 100  0.058830804  2.497569779

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  1.26870920  1.60886343
#> 2   002  0.39983632 -0.78564210
#> 3   003 -1.46173818  0.28026643
#> 4   004 -0.31149584  0.83456581
#> 5   005 -0.40098696  0.59765588
#> 6   006  1.38365347  1.06655463
#> 7   007 -1.37531281 -1.02145106
#> 8   008  1.01653546  0.61778506
#> 9   009 -0.45295403 -0.93469010
#> 10  010 -0.15134264  0.19828292
#> 11  011 -0.07584009 -0.82800800
#> 12  012  1.40107864 -0.23806767
#> 13  013 -0.13050494  0.07758468
#> 14  014 -0.62907461  0.79106234
#> 15  015 -0.91556565  1.13914382
#> 16  016  0.94417755  0.77872388
#> 17  017  0.95015020  0.82052425
#> 18  018 -0.72597989 -2.15886556
#> 19  019 -0.56020656  0.09835011
#> 20  020 -0.26233763 -0.23124328
#> 21  021 -1.07901466 -0.79966503
#> 22  022  0.76271531  0.39369856
#> 23  023  0.11904533  0.45789717
#> 24  024  1.05731211 -0.75009446
#> 25  025  1.39712256  0.47668384
#> 26  026  0.67421791 -1.71948224
#> 27  027  0.20991189  0.26369525
#> 28  028  1.42829270 -0.26718198
#> 29  029  0.57046046  1.43039714
#> 30  030  1.33266602  0.64542633
#> 31  031  1.43570981  0.57978625
#> 32  032  0.38197890 -0.34547585
#> 33  033  0.61051661  0.46202259
#> 34  034  0.20344983 -0.13838712
#> 35  035 -0.53926441 -0.24119088
#> 36  036  0.21293097 -0.67208871
#> 37  037  1.07090792 -0.41973254
#> 38  038 -0.74374306  0.58495730
#> 39  039 -0.66561206 -1.95310315
#> 40  040  1.17263761  0.35404187
#> 41  041  0.73957139  0.38204759
#> 42  042 -2.13654173 -0.42350946
#> 43  043 -1.57252824 -0.20247934
#> 44  044 -0.21079168  0.27982934
#> 45  045  0.62351397 -2.00190467
#> 46  046  0.20223055  0.56904193
#> 47  047 -0.64912027 -1.63167502
#> 48  048 -1.12185949 -0.77531984
#> 49  049 -0.44605448 -0.63907265
#> 50  050 -1.85353516  0.76666353
#> 51  051 -0.83004643 -0.82113625
#> 52  052 -1.06678814  0.59351314
#> 53  053 -1.30084686 -0.76810204
#> 54  054  0.20825584  1.70072921
#> 55  055  2.50118245  1.65418155
#> 56  056  0.20290481  0.42947976
#> 57  057  0.29052799 -0.71311336
#> 58  058 -0.18076522  0.63902843
#> 59  059  0.52812789  1.66119102
#> 60  060  0.50402198 -0.38927927
#> 61  061  0.20680816 -0.17365597
#> 62  062 -1.03799232  0.13398070
#> 63  063  0.28333329 -0.73552722
#> 64  064 -2.16877826 -0.14278315
#> 65  065  1.31939221  0.57991538
#> 66  066  0.94078371  0.19266095
#> 67  067 -0.11264882 -1.00751594
#> 68  068 -1.09627132 -0.43465366
#> 69  069  0.22163863 -0.83201385
#> 70  070 -1.01423341 -1.34807885
#> 71  071 -0.12573286  1.81854832
#> 72  072 -0.39188307 -0.72259163
#> 73  073  0.36489355 -0.09355900
#> 74  074  0.53079712  1.28574740
#> 75  075 -0.22995615  0.55309696
#> 76  076 -0.14927801  1.19580117
#> 77  077 -0.51908678  0.11260717
#> 78  078 -0.18043105 -0.39575588
#> 79  079 -1.44858407 -0.89316487
#> 80  080 -1.08609804 -0.20943414
#> 81  081  1.58527683 -0.24915438
#> 82  082 -1.08816083 -0.69529058
#> 83  083  0.11481699  1.99444670
#> 84  084  0.16263904 -0.74651813
#> 85  085  1.25869287  1.62575691
#> 86  086  0.02967447  0.89049121
#> 87  087 -0.17835631 -0.47534819
#> 88  088  1.95653251  0.29673157
#> 89  089 -0.41893254 -2.25735335
#> 90  090 -0.44308161 -0.91024283
#> 91  091  0.53506693  1.31648838
#> 92  092 -0.03973022 -0.08689496
#> 93  093  0.20756972 -1.87734466
#> 94  094 -1.56087608 -0.12824933
#> 95  095 -0.19784016  0.43021753
#> 96  096 -0.76147276 -0.59213929
#> 97  097 -0.47049274 -0.19302293
#> 98  098 -0.91970916 -0.20959012
#> 99  099  0.17122188  0.39922657
#> 100 100  0.35083448  1.33903346