Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4861014
#> Y_2 0.4861014 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.5217495 -0.9500968
#> 2  1.1997574 -0.8048933
#> 3 -1.0394213 -0.2137789
#> 4  0.8302429  1.4194412
#> 5  0.4786754  0.9826432

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1         Y_2
#> 1   001 -1.269130083 -0.60937310
#> 2   002 -1.208581876 -0.74373406
#> 3   003 -0.804706424  0.25377055
#> 4   004  0.383637160  0.10745895
#> 5   005 -0.303680696  0.40721516
#> 6   006 -0.684204082 -0.61188567
#> 7   007 -1.511637426 -0.85439055
#> 8   008  0.453996088 -0.72792626
#> 9   009  1.741780814  0.85758616
#> 10  010  0.335584415  1.01139448
#> 11  011 -0.443236898 -0.78265441
#> 12  012  0.662594781 -0.64095148
#> 13  013  0.210528812  0.70647098
#> 14  014  0.426647594  2.33054550
#> 15  015  0.740207392  0.53697826
#> 16  016 -0.861582410  0.18003817
#> 17  017 -0.736590960  0.05058056
#> 18  018  1.682748357  1.42779766
#> 19  019 -1.435294140 -0.01175789
#> 20  020  2.537028249  0.47457140
#> 21  021 -0.014705753 -1.12255344
#> 22  022 -1.264132426  0.09171053
#> 23  023 -0.747432637 -0.86538870
#> 24  024  1.031742039  0.56388517
#> 25  025  0.339859925  0.48683348
#> 26  026  2.348728919  0.96126460
#> 27  027 -1.115049666  1.13179528
#> 28  028  0.280485186 -0.19752699
#> 29  029  0.077583407 -0.21184387
#> 30  030  0.895892022 -0.09983981
#> 31  031  0.429034213  1.83522989
#> 32  032 -0.523552895  0.90467209
#> 33  033  2.027616619  0.80029936
#> 34  034 -0.612935910  0.85533275
#> 35  035  1.854863576  1.59501400
#> 36  036  1.371757550  0.65129303
#> 37  037  0.752691583  0.27772072
#> 38  038  0.127600781 -0.48088648
#> 39  039  0.828688775 -0.20476771
#> 40  040 -0.406101807  0.70325055
#> 41  041 -0.469831840  0.13078618
#> 42  042 -0.318764473  1.50470644
#> 43  043 -0.502483243  0.27025121
#> 44  044  0.444819683 -0.17562481
#> 45  045  0.488917182  1.09030648
#> 46  046  0.006236948 -0.52343045
#> 47  047  0.796532673  0.21141531
#> 48  048  0.295599210 -0.85291168
#> 49  049 -0.044032142 -0.67607533
#> 50  050 -1.150818612 -0.61162625
#> 51  051 -1.192607521 -2.55152435
#> 52  052 -0.487922454  0.26675365
#> 53  053  1.538715879  2.99635493
#> 54  054 -0.011992643 -0.15419638
#> 55  055 -0.456675470 -0.07927582
#> 56  056 -0.183710813  0.58715119
#> 57  057  0.967977252  0.16091035
#> 58  058  0.219689829 -1.09317762
#> 59  059  2.014970964  3.68722479
#> 60  060  0.217336978  0.05476591
#> 61  061  0.357038751 -0.80528674
#> 62  062  1.062993214 -1.13886252
#> 63  063  0.299806254  0.15153654
#> 64  064 -0.297857022 -0.29031894
#> 65  065  1.406852691  1.51046243
#> 66  066 -1.386869883 -3.19697439
#> 67  067  0.383505107 -1.11555429
#> 68  068  0.234315894 -2.21599821
#> 69  069 -0.187029766 -0.20859387
#> 70  070 -0.454717056 -0.85009082
#> 71  071 -0.265796398 -0.12625017
#> 72  072 -0.059201676  0.03970130
#> 73  073  0.134190854  1.23612156
#> 74  074  0.168208317 -1.19696361
#> 75  075  1.498750756  1.15068821
#> 76  076 -1.728501475  0.11894484
#> 77  077  0.551595571 -0.29714457
#> 78  078 -0.775542392 -0.55690439
#> 79  079  1.351429030  0.60491957
#> 80  080  0.796852419  1.16283403
#> 81  081 -1.687030892 -0.20213749
#> 82  082  0.791113216 -0.29455088
#> 83  083  0.748746945  1.36883711
#> 84  084  0.100959634 -0.28824679
#> 85  085 -0.063092770 -0.27703621
#> 86  086  0.543897031  0.11952079
#> 87  087  1.539822347  1.15869633
#> 88  088 -0.662870103 -1.79823350
#> 89  089  0.356645685  1.22903864
#> 90  090  0.635886000  0.84420630
#> 91  091 -0.206184313  1.58938128
#> 92  092  1.431568452  0.69874253
#> 93  093 -0.004438730  1.27637866
#> 94  094  1.576690452  1.94652289
#> 95  095 -0.024966040  0.02879236
#> 96  096 -0.939110888  0.47032560
#> 97  097  0.905174958 -1.03673538
#> 98  098 -0.234458179 -1.07069602
#> 99  099 -0.446335139 -0.32336999
#> 100 100  1.620343540  1.83988387

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.95131827 -0.04676717
#> 2   002 -0.98034428 -1.58919767
#> 3   003 -1.27388804  0.93981604
#> 4   004 -0.57179895 -0.41206478
#> 5   005  0.11182103  0.39367459
#> 6   006  0.84167403  0.06369032
#> 7   007  0.22508279  2.19846395
#> 8   008  0.04015313  0.38600677
#> 9   009 -1.67682622 -1.50323304
#> 10  010  0.75080764  0.40151040
#> 11  011 -1.18793279  0.40950982
#> 12  012  0.26446937  0.71218849
#> 13  013 -0.73375452 -0.82042116
#> 14  014 -0.95815588  1.43443212
#> 15  015 -2.46464086 -1.25451249
#> 16  016 -0.32677641 -1.08886801
#> 17  017 -0.97469588 -0.86500296
#> 18  018  0.08398010 -1.32387496
#> 19  019 -0.40491629  1.05167482
#> 20  020 -1.84153233 -0.12856039
#> 21  021 -0.70420345 -1.22472067
#> 22  022  2.25394649  2.98363430
#> 23  023  0.22625036  0.36392561
#> 24  024 -0.88646051  1.42414876
#> 25  025  2.13734459  0.40586419
#> 26  026 -0.23768550  1.42947191
#> 27  027 -0.37622994  0.11378396
#> 28  028  0.36534273  1.38798862
#> 29  029  0.79405636  0.33111270
#> 30  030 -1.44232305 -0.65455652
#> 31  031  0.95685364  0.57898217
#> 32  032 -0.43074571 -1.67202023
#> 33  033  1.19205289  1.12451458
#> 34  034 -0.85050871 -2.43958280
#> 35  035 -0.69610114  1.22834418
#> 36  036 -0.53993633  0.03527408
#> 37  037 -1.44267529 -0.70411125
#> 38  038 -0.41026068 -0.41509203
#> 39  039 -0.28518380 -0.57437881
#> 40  040  0.29034610  0.71910396
#> 41  041 -1.03051709  0.60179111
#> 42  042 -0.75859069 -1.87348252
#> 43  043 -0.08512410 -0.08273327
#> 44  044  0.12744383  0.64369129
#> 45  045 -0.86529274 -1.43787329
#> 46  046  0.25244972 -0.56995187
#> 47  047 -0.33217029  0.94176987
#> 48  048 -1.10901006 -0.85588630
#> 49  049  2.00789528  1.31870700
#> 50  050 -0.80448441 -0.63918622
#> 51  051 -0.05182727 -0.43249722
#> 52  052  0.08690216  0.17289836
#> 53  053 -0.85905472 -0.34070988
#> 54  054  0.35711259 -0.06259454
#> 55  055  1.12603905  0.83750580
#> 56  056 -0.93486717 -1.26821269
#> 57  057 -0.46592946  0.23852430
#> 58  058 -0.46586384  0.08177129
#> 59  059  0.19490939 -0.75628760
#> 60  060  1.70538811 -0.66014808
#> 61  061 -1.59884834 -0.05056381
#> 62  062  1.92243715  0.81825433
#> 63  063 -0.87014528 -0.44684414
#> 64  064  0.08226905  0.25996589
#> 65  065  0.48813160 -0.24937664
#> 66  066  0.45158538 -0.49554651
#> 67  067  1.79311963  0.74934111
#> 68  068 -0.78730303  0.12127744
#> 69  069 -0.03481624  0.09942017
#> 70  070  0.99053756  1.11834368
#> 71  071 -0.14697582  0.14435800
#> 72  072  2.25395592 -0.38668794
#> 73  073 -1.03452005  0.15950152
#> 74  074  0.37577790 -1.15969325
#> 75  075 -1.67901003  0.06431982
#> 76  076 -0.10293802 -0.13764482
#> 77  077 -0.80538596 -0.22108226
#> 78  078 -0.66977914 -0.08593293
#> 79  079  1.20049511  0.98908058
#> 80  080 -0.15770650  0.62035603
#> 81  081 -0.12784543  0.03156952
#> 82  082  0.77947673  1.08368764
#> 83  083 -1.21175105 -1.14686501
#> 84  084 -0.88496062 -0.99705418
#> 85  085  0.27413687  0.55713011
#> 86  086  0.56345701  1.64164568
#> 87  087  1.69756745  0.30174200
#> 88  088  0.90336755  1.43756302
#> 89  089  0.13215409  0.90629594
#> 90  090  0.85596314  1.32395515
#> 91  091 -2.55795358 -0.96591164
#> 92  092 -0.20772377 -0.84254414
#> 93  093  0.70777048  2.90045817
#> 94  094  1.60041242 -0.86731088
#> 95  095  0.75224353  0.08659644
#> 96  096  0.80623238 -0.97554324
#> 97  097  0.23609347  0.26671148
#> 98  098  1.01582260  0.18227755
#> 99  099  2.44887359  3.08177778
#> 100 100  0.04691366 -0.47780322