Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4604059
#> Y_2 0.4604059 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -0.71583543 -0.9141902
#> 2 -1.06072043 -0.5712936
#> 3 -1.76954450  0.5656883
#> 4  0.20143078 -0.2841560
#> 5  0.08581595  0.9192520

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -0.01609155  0.36724386
#> 2   002  1.64255713  0.91757327
#> 3   003  0.39114644  0.54946431
#> 4   004  1.12275828  1.68876414
#> 5   005  0.42395640 -0.59747978
#> 6   006 -0.88319321  0.10385615
#> 7   007  1.02926976  0.36202415
#> 8   008  0.25020292  1.04019576
#> 9   009  1.32401640  0.86467437
#> 10  010 -1.54151255 -1.34856439
#> 11  011 -0.51604962  0.89364254
#> 12  012 -0.81179417  0.58808613
#> 13  013 -1.86307940  0.48634369
#> 14  014  0.53558010  0.50027394
#> 15  015 -1.26453352  0.99779919
#> 16  016  0.07369663 -0.56001463
#> 17  017 -0.90127046 -1.33987993
#> 18  018 -2.08747723 -1.32464268
#> 19  019 -0.51757817 -0.80053386
#> 20  020  0.04285575 -0.06394651
#> 21  021 -0.28194920  0.40065098
#> 22  022 -0.91042760 -0.77297720
#> 23  023 -0.27948196 -1.49728088
#> 24  024 -0.34770737  0.79616001
#> 25  025  1.23308724  0.75059001
#> 26  026 -0.93363924 -0.43472269
#> 27  027 -0.17604466  0.07632156
#> 28  028  0.49478775 -0.15950383
#> 29  029 -0.69778924 -1.15529271
#> 30  030 -0.05117309  0.67367714
#> 31  031 -0.97420023 -1.14980912
#> 32  032 -1.71195205 -1.48046381
#> 33  033 -0.33744190  0.03050267
#> 34  034  0.12512381  0.67619648
#> 35  035 -0.48350549  0.56063992
#> 36  036  2.04465514  0.54277508
#> 37  037  1.00767971  1.51825344
#> 38  038 -0.05143364 -0.93037796
#> 39  039 -0.06033939  0.22739522
#> 40  040 -2.17163504 -0.81214291
#> 41  041 -1.25437645  0.28730483
#> 42  042  1.93397128  1.65257565
#> 43  043  0.13898183 -0.61091691
#> 44  044  0.33855427  0.86779517
#> 45  045  0.55635813  0.42829347
#> 46  046 -1.26758617 -0.99413413
#> 47  047 -0.44799680 -0.30803848
#> 48  048  1.60899963  1.79934074
#> 49  049  0.78944188 -0.91090925
#> 50  050 -1.02135081 -1.34707469
#> 51  051 -1.13668489 -0.04277016
#> 52  052 -0.14092536  0.35989629
#> 53  053  1.49733003  2.31187974
#> 54  054 -0.61029114 -1.00968767
#> 55  055  0.06347648 -0.95186063
#> 56  056  0.53826200  0.14236813
#> 57  057  0.26670415  0.30869329
#> 58  058 -1.95367134 -3.21038064
#> 59  059  0.83446413 -1.41434357
#> 60  060 -3.00751288 -1.50993427
#> 61  061  1.17313122 -1.05556709
#> 62  062 -0.40284461 -0.77638453
#> 63  063 -1.61947768 -0.36561855
#> 64  064  0.55135760  0.51202093
#> 65  065  0.32484418  0.68448718
#> 66  066  1.12588349  0.24308525
#> 67  067  0.35643592  0.06890267
#> 68  068 -0.30291216 -1.48080673
#> 69  069  0.75779164  1.55100244
#> 70  070 -0.51582412 -0.28737213
#> 71  071 -1.15666033 -2.05564640
#> 72  072 -0.20514965 -0.37733772
#> 73  073 -0.99149660 -0.24850591
#> 74  074 -0.56212692 -0.71863831
#> 75  075  0.81942139  1.42641152
#> 76  076  0.22598112  0.44373418
#> 77  077 -0.08691164 -2.79402528
#> 78  078 -0.88793148  0.85560732
#> 79  079  0.87563118 -0.28015566
#> 80  080 -0.10140612 -0.36589361
#> 81  081 -0.94564402 -0.85065208
#> 82  082 -1.25655524 -0.63533797
#> 83  083 -0.76570107  1.01293836
#> 84  084 -1.64667991 -0.99335552
#> 85  085  2.19672024  1.93511650
#> 86  086  1.78385253  0.64939088
#> 87  087  0.62968894  0.51147360
#> 88  088 -1.15965412 -1.30812979
#> 89  089  0.52906814 -0.23159819
#> 90  090 -0.42932500 -1.90556655
#> 91  091 -1.13251569 -1.14806337
#> 92  092 -1.01469070 -1.13245195
#> 93  093  0.15217082 -1.59437523
#> 94  094 -0.14032886 -0.90172746
#> 95  095  0.07556020  0.71956490
#> 96  096  0.47727134  0.04593594
#> 97  097  2.31389257  2.32307721
#> 98  098 -0.98842982 -0.54562934
#> 99  099  0.41860910  0.06296800
#> 100 100 -0.94792854 -1.51871543

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  1.65268344  2.22252960
#> 2   002  0.25195197  0.15582765
#> 3   003  0.57126838 -0.04294546
#> 4   004 -1.48680146 -0.75352596
#> 5   005  0.76007325 -0.59979202
#> 6   006  0.53367822  2.11038874
#> 7   007 -0.08748734  0.53869137
#> 8   008 -0.65087258 -0.69472177
#> 9   009  0.86168803 -0.64472436
#> 10  010 -0.28313773 -0.77825600
#> 11  011 -0.11586840  0.30033137
#> 12  012  1.20359560  0.06719261
#> 13  013 -0.46759866 -0.57617189
#> 14  014 -0.83305759 -0.97172593
#> 15  015 -0.68802390  0.32098466
#> 16  016 -0.20009415  0.49973920
#> 17  017  1.94786657  0.17006634
#> 18  018 -0.48467359 -0.15254268
#> 19  019 -0.11299362  0.90396501
#> 20  020  0.93004061  1.09204863
#> 21  021 -0.81732367 -0.64888757
#> 22  022  0.78300983 -0.52124357
#> 23  023  0.23509944  0.84697029
#> 24  024 -1.87323581  0.51102999
#> 25  025 -0.45589279  0.61931742
#> 26  026 -1.99444644 -2.57716954
#> 27  027  0.90943146  0.91093160
#> 28  028 -0.38436613  0.62392218
#> 29  029  1.54121282  0.83849055
#> 30  030  0.45537995 -0.10548242
#> 31  031  0.37737148 -0.26140508
#> 32  032 -0.55027238 -0.08152565
#> 33  033 -0.36116590 -1.70451136
#> 34  034 -0.45748570  0.65919534
#> 35  035  1.29184108  0.85257551
#> 36  036 -0.93959230  0.50511679
#> 37  037  1.37444592  0.50088875
#> 38  038 -0.70360642 -0.73298814
#> 39  039  1.60857392  1.03613920
#> 40  040  0.90755962  0.01808012
#> 41  041  0.53145624  0.19832998
#> 42  042 -0.16701265  0.74931416
#> 43  043 -0.29219904  1.43142971
#> 44  044 -0.99779617 -0.80204572
#> 45  045 -0.07761628  0.34296541
#> 46  046 -1.57298416  0.16848345
#> 47  047  0.80011240  1.47993729
#> 48  048  0.20240676 -1.34280515
#> 49  049 -0.09710820  0.15998583
#> 50  050 -1.90520619 -0.63439848
#> 51  051  0.50188462  0.83538439
#> 52  052  1.15058893  1.29050047
#> 53  053 -0.52206776 -0.58245382
#> 54  054  0.71458449  0.39745906
#> 55  055 -0.83867343 -0.89807455
#> 56  056 -1.05790118  0.00201498
#> 57  057 -0.63914670 -0.45941062
#> 58  058 -1.37972606 -0.97190919
#> 59  059 -1.62764831 -2.35351072
#> 60  060  0.65154233 -1.11095915
#> 61  061  0.85359323  0.82821276
#> 62  062 -0.34482345 -0.81924873
#> 63  063  1.84227922  0.50660568
#> 64  064  1.91355669  1.86398894
#> 65  065 -1.55818842 -0.07699367
#> 66  066 -0.57019549  0.09592453
#> 67  067  0.64757555 -0.68164509
#> 68  068 -0.28295781  0.69861649
#> 69  069 -0.64926872 -0.77709449
#> 70  070 -0.78710799 -1.13948542
#> 71  071 -2.00898823  0.24496147
#> 72  072 -0.31014093  0.27097508
#> 73  073  1.49573283  0.41818016
#> 74  074  0.06421502 -0.33793475
#> 75  075 -1.07312431 -1.59226174
#> 76  076  0.09864522  0.18463017
#> 77  077 -1.30971531 -0.23920481
#> 78  078  2.03153818 -0.53625369
#> 79  079  0.61824888  0.52461000
#> 80  080 -1.38714106 -0.14758100
#> 81  081  0.25917116 -0.29481749
#> 82  082  0.50710852  0.54246099
#> 83  083  0.39877396  0.39162812
#> 84  084  1.91574725  0.34836416
#> 85  085 -0.53197043 -0.68911516
#> 86  086  0.85273898  0.40865560
#> 87  087  1.58115469  0.22900808
#> 88  088  2.49372801  0.61957690
#> 89  089 -1.23983532  0.34605899
#> 90  090  0.74843946  1.27606907
#> 91  091  0.61222582  1.09465846
#> 92  092 -1.65402929 -0.88408549
#> 93  093 -1.25470818  0.25454132
#> 94  094 -1.17633085 -1.84082715
#> 95  095  2.48175450  1.36492374
#> 96  096  1.02727351  2.86528799
#> 97  097 -0.73231720 -1.45124222
#> 98  098 -1.54720527 -0.31075290
#> 99  099 -0.93354564 -1.77499292
#> 100 100 -0.51649454  0.67614429