Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5021756
#> Y_2 0.5021756 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.3034767 -0.5741177
#> 2  0.4982123 -0.8064091
#> 3 -2.0436541 -0.8915104
#> 4 -0.3243586 -2.1100092
#> 5  1.0351673  0.4563483

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.593884309 -0.020205634
#> 2   002 -3.393881915 -0.766813276
#> 3   003  1.310166446  0.295301661
#> 4   004 -0.184301310 -0.686274720
#> 5   005  0.167024538  0.906723569
#> 6   006 -1.927909572 -1.996094115
#> 7   007  0.226520740 -0.284514664
#> 8   008 -0.507945118 -0.952866034
#> 9   009 -0.483660832 -0.758917032
#> 10  010 -0.801254037 -1.079526734
#> 11  011  0.340814572  1.149760798
#> 12  012 -0.874361246 -0.727844229
#> 13  013  0.724169764 -0.356365373
#> 14  014  0.005616201  0.798659188
#> 15  015  1.540736326  0.005338556
#> 16  016 -1.471955126 -0.101193550
#> 17  017 -0.321726833 -0.830050106
#> 18  018 -0.324266138 -0.645846146
#> 19  019  0.730841242  0.581185603
#> 20  020  0.258559639 -0.198144413
#> 21  021  0.161930924  0.757363002
#> 22  022 -1.265774538 -0.913271703
#> 23  023 -0.917705091 -0.978509561
#> 24  024  0.250832052 -0.328338366
#> 25  025  0.382162595  0.331672723
#> 26  026 -0.086213166  0.711143842
#> 27  027 -0.006228374 -0.410180171
#> 28  028  0.584989587 -0.099026830
#> 29  029  0.972296255  0.424555783
#> 30  030 -0.712408817 -1.206474625
#> 31  031  0.277058868  1.074171698
#> 32  032 -0.444323744 -0.123874730
#> 33  033  0.568402710  0.651802381
#> 34  034 -0.621740966 -0.680301555
#> 35  035  0.300855369 -0.286747465
#> 36  036  0.661071564  0.365419408
#> 37  037 -1.212046893 -1.121498851
#> 38  038 -0.243000744  0.808333317
#> 39  039 -0.156498747 -0.174158310
#> 40  040  1.369197363  1.758775343
#> 41  041  1.755849979 -0.053843802
#> 42  042 -0.169234682 -0.255238298
#> 43  043 -1.814660797 -1.627027833
#> 44  044  0.896377943 -0.280484017
#> 45  045  0.651125873 -0.367209644
#> 46  046 -0.545477671 -2.012049198
#> 47  047  0.945790481 -0.142649295
#> 48  048 -0.371590550 -2.128923266
#> 49  049 -1.892683671  1.160349407
#> 50  050  1.218533083  0.165149865
#> 51  051 -0.272621977 -0.567332864
#> 52  052 -1.291650695 -0.439213812
#> 53  053  1.979518397  1.440104678
#> 54  054 -0.052379989 -0.649548161
#> 55  055 -1.221458979 -0.818595404
#> 56  056  2.532280312  1.688586618
#> 57  057 -0.757597802 -0.807896538
#> 58  058  1.076213365  0.959184063
#> 59  059 -0.903900103 -1.012744705
#> 60  060 -1.132332264 -0.986399543
#> 61  061 -0.477473030  0.269376307
#> 62  062 -0.715590423 -1.575224188
#> 63  063  1.130923927 -0.293416286
#> 64  064  0.086553290  0.543000073
#> 65  065 -0.123599515 -0.119943315
#> 66  066 -0.350925896 -1.200986482
#> 67  067  0.744181045  0.424606796
#> 68  068 -0.429259021 -0.463862090
#> 69  069  0.770769466  0.758822011
#> 70  070  1.620970133 -0.555362115
#> 71  071 -0.500887966  1.185194217
#> 72  072  0.734992671 -0.284194018
#> 73  073 -0.181689333 -0.461344145
#> 74  074  0.296943029 -1.053185589
#> 75  075 -1.517135556  0.412462011
#> 76  076  0.116800651  1.545456596
#> 77  077  1.048650645  0.301689132
#> 78  078  0.375855337  1.354581587
#> 79  079  1.297364352 -0.010401785
#> 80  080  0.801571664  0.709598143
#> 81  081 -0.670646079 -1.330215216
#> 82  082  1.062809871 -0.125282419
#> 83  083 -1.685623247 -0.129039860
#> 84  084 -0.859493475 -1.016627774
#> 85  085 -0.470227693  0.743810350
#> 86  086 -0.817373100  0.749019337
#> 87  087 -1.160216382  0.549211108
#> 88  088 -0.045241064  0.243694487
#> 89  089 -0.487765790 -1.422200782
#> 90  090 -0.202344622 -0.415631807
#> 91  091  1.412853719  1.051061599
#> 92  092  0.078696286  0.191658752
#> 93  093 -0.534669563  1.111951555
#> 94  094  0.116894391 -1.495759927
#> 95  095 -0.875431433 -1.841680860
#> 96  096 -0.476795732  1.913430765
#> 97  097  1.554177331  1.061979772
#> 98  098  0.380924144  2.810962293
#> 99  099  0.394430736  1.207796765
#> 100 100 -0.011865389  0.561512113

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.897336606  0.13998116
#> 2   002  0.513077841  0.14477180
#> 3   003 -1.023932602  0.25019500
#> 4   004 -0.809858655 -1.02492268
#> 5   005 -0.701874317 -1.05152743
#> 6   006  0.282831652  0.36358600
#> 7   007 -0.518209787 -0.44430174
#> 8   008  0.391892050  1.02974732
#> 9   009  1.119008851 -0.42838760
#> 10  010  0.901835246 -0.63793744
#> 11  011  1.077150465  0.25554320
#> 12  012  0.306019055 -1.19458730
#> 13  013 -0.156159712  0.13425306
#> 14  014  0.917805808 -0.10182743
#> 15  015 -0.929366867 -0.29391821
#> 16  016  0.248885398  0.21829532
#> 17  017 -3.301143918 -2.69214199
#> 18  018  1.388369716  0.60748877
#> 19  019  0.482856900  0.33693066
#> 20  020  0.013544255 -1.77416715
#> 21  021 -1.246928882 -1.64682799
#> 22  022  1.538488021 -0.61862619
#> 23  023 -0.549637849 -0.45182353
#> 24  024 -1.769885040  0.24367112
#> 25  025  0.763895075  0.86546985
#> 26  026  1.480941016  2.16984194
#> 27  027  2.137519183  1.56404249
#> 28  028 -2.334478728 -2.57309726
#> 29  029  1.862984527  0.22990138
#> 30  030 -1.419288972 -0.34540719
#> 31  031  0.192330817  1.34573606
#> 32  032  0.028199831  0.60568867
#> 33  033 -0.872748254 -0.19356930
#> 34  034  0.094777016  0.19666348
#> 35  035 -1.234467685 -0.46552646
#> 36  036 -0.841831015 -0.42270972
#> 37  037 -1.428459621 -1.37179595
#> 38  038 -0.767376632 -1.21552061
#> 39  039  1.386814407 -0.77412946
#> 40  040 -0.387730763  0.82164273
#> 41  041  0.422990312  1.43354541
#> 42  042 -0.914600499 -1.82567923
#> 43  043  0.358664022 -0.07054180
#> 44  044 -0.888752149 -0.82105578
#> 45  045 -0.932616154 -0.42932392
#> 46  046 -0.262530920  1.16232407
#> 47  047 -0.575667445  0.36601698
#> 48  048 -0.923948263 -2.45859469
#> 49  049  0.014032375 -0.25602501
#> 50  050 -1.141442741  0.27537038
#> 51  051 -0.661404848  0.79829506
#> 52  052  1.492583067  0.97902800
#> 53  053 -1.845398425 -1.91256809
#> 54  054 -0.266075551 -1.91754846
#> 55  055  1.254978957 -0.16651959
#> 56  056 -3.061765087 -2.98336881
#> 57  057 -2.003545700 -1.29140805
#> 58  058 -0.335506542  0.41351374
#> 59  059 -0.488538653 -1.68678632
#> 60  060  0.009657606 -1.05918835
#> 61  061 -1.265600789 -0.33848987
#> 62  062  0.442291663  0.28096185
#> 63  063  0.007827752  2.02244352
#> 64  064 -0.959980521 -1.10889240
#> 65  065 -0.432331492  1.68424202
#> 66  066 -0.243103905 -0.22031379
#> 67  067  1.036105440  0.19514351
#> 68  068  2.847492107  1.94220280
#> 69  069 -0.002688279 -0.68844810
#> 70  070  0.431306497  1.08112852
#> 71  071 -0.017605276  1.26191561
#> 72  072  1.254077851  1.13262991
#> 73  073  0.013734259 -0.39207433
#> 74  074  1.338403052 -0.67731728
#> 75  075 -1.115483645 -0.28694991
#> 76  076 -0.767650418  1.23422227
#> 77  077  1.414557140  1.08798643
#> 78  078 -0.002594415  0.35296401
#> 79  079 -0.240340478 -0.35704977
#> 80  080  0.937497871  0.64259777
#> 81  081  0.953003211  0.04000399
#> 82  082 -0.302270738  0.02319658
#> 83  083  0.714001092 -0.65725942
#> 84  084 -1.658954648 -0.55926351
#> 85  085 -0.394963009 -1.45459818
#> 86  086  0.983887548  0.64647995
#> 87  087  0.922334447  0.99108540
#> 88  088  0.484479166 -0.70434046
#> 89  089 -0.366101614  0.10994581
#> 90  090 -0.212105795 -1.26925107
#> 91  091 -0.289353255 -0.67282610
#> 92  092 -0.525337655 -0.40604209
#> 93  093  0.250270375  0.14253799
#> 94  094 -0.034015459 -0.04718581
#> 95  095  0.730997527 -0.73152826
#> 96  096 -0.066538204 -0.81084123
#> 97  097 -0.812327133 -0.55715104
#> 98  098 -1.394337434 -1.89837881
#> 99  099 -0.637568001 -1.11568794
#> 100 100 -0.324603604 -0.51632948