Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5292943
#> Y_2 0.5292943 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1          Y_2
#> 1  0.09940397  0.079901316
#> 2 -0.94504725  0.865353536
#> 3 -0.17829502  0.195334912
#> 4  0.38273099  0.002304203
#> 5 -0.47210968 -0.329122657

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001 -1.41507707  0.23694069
#> 2   002  1.51280960  1.78111987
#> 3   003 -0.16476649  0.29360117
#> 4   004  0.67660506  0.22566963
#> 5   005  0.81771289  0.77947712
#> 6   006  1.00103805  2.31499174
#> 7   007  0.10873043  0.21549415
#> 8   008 -0.53537368  1.88457967
#> 9   009 -1.49086878 -0.92111508
#> 10  010 -0.70322484 -0.52749133
#> 11  011 -1.39804431 -1.26782135
#> 12  012  0.46823046  1.18518227
#> 13  013 -0.19135514 -0.56202639
#> 14  014  0.58344930  0.21560423
#> 15  015 -1.85850470 -0.14289019
#> 16  016  0.63357201  0.42510744
#> 17  017  1.02415280  0.46981194
#> 18  018 -0.15021301 -1.49078715
#> 19  019 -0.15496234 -0.13059617
#> 20  020  0.60963851 -0.81740445
#> 21  021  0.24020430 -0.82826380
#> 22  022 -0.52148993 -1.45160874
#> 23  023  1.77590282  0.91730180
#> 24  024  0.07661705  1.60246961
#> 25  025 -0.24509342  0.09454389
#> 26  026  1.96167856 -0.27605750
#> 27  027  1.53544186  0.98454880
#> 28  028  0.23423969 -0.43934111
#> 29  029  0.18534445  1.18293408
#> 30  030  0.87551176  0.94764156
#> 31  031  0.05265508 -1.38011577
#> 32  032 -0.20883876 -0.40895419
#> 33  033 -0.66277436  0.32248875
#> 34  034 -1.59212238 -1.17828060
#> 35  035  2.27405932  0.71428947
#> 36  036  1.61569444  1.11828490
#> 37  037 -0.51272256 -0.77942206
#> 38  038  0.56598536 -0.04306941
#> 39  039 -1.31194238 -0.42800046
#> 40  040  1.17138249  1.21508698
#> 41  041 -0.85369225  0.93191885
#> 42  042 -0.05792010 -0.20772147
#> 43  043  0.97228168  1.09914075
#> 44  044 -1.54956650  0.53333239
#> 45  045 -0.77769230  0.69808580
#> 46  046  0.89323300  2.06107845
#> 47  047 -0.62199136 -0.08558276
#> 48  048 -1.41596261 -1.28011179
#> 49  049 -0.31424206  0.25153338
#> 50  050  0.11606988 -0.74098880
#> 51  051  0.02182382 -0.51637564
#> 52  052 -0.67167379 -0.91076220
#> 53  053 -1.06365515 -0.81427941
#> 54  054  0.41699768  0.41535995
#> 55  055  1.62528450 -1.15296115
#> 56  056  0.68027133 -0.42622469
#> 57  057  1.31036551  1.50733705
#> 58  058  0.23444208  1.23723611
#> 59  059 -0.58093880  0.13349932
#> 60  060  1.13225356  0.27459965
#> 61  061  0.07087536  1.21508799
#> 62  062 -1.23787313 -0.25387105
#> 63  063  0.42068787  1.53683423
#> 64  064  1.48668491  2.31342039
#> 65  065  0.70255524 -0.51971124
#> 66  066 -1.11482526  0.55560255
#> 67  067 -0.67530137 -0.83031412
#> 68  068  1.53116561 -0.18199638
#> 69  069 -0.17620168 -0.37277850
#> 70  070  0.66450747 -0.04816166
#> 71  071 -0.49764715 -1.11720720
#> 72  072 -0.10563495  0.28291283
#> 73  073 -0.41795234 -0.48903939
#> 74  074 -0.17587428  0.23861288
#> 75  075 -0.26915187 -1.16765417
#> 76  076  0.77639250 -0.67549529
#> 77  077  1.33196482  1.66211306
#> 78  078  0.36487036 -0.49058314
#> 79  079  0.59472930  0.27022492
#> 80  080 -1.51986616 -0.39977516
#> 81  081  0.13100610  0.41013638
#> 82  082 -0.16994573 -0.99098450
#> 83  083  0.16593866  1.21437358
#> 84  084  1.05355058 -0.35942348
#> 85  085 -0.90335955 -0.30394176
#> 86  086  0.25959041 -0.96427905
#> 87  087  2.11064836  1.78313832
#> 88  088 -0.63365091 -0.22418557
#> 89  089 -0.65040014  0.63899983
#> 90  090  1.48682311  1.90440156
#> 91  091  0.17608085 -0.87159312
#> 92  092  0.95717236  0.90636356
#> 93  093  0.50532610 -1.36208518
#> 94  094  0.43558556  1.15133309
#> 95  095 -1.53886557 -0.98010344
#> 96  096  0.32829164 -0.76239061
#> 97  097  1.22634191  0.42652754
#> 98  098  1.11077145  0.71207463
#> 99  099  1.08892729  1.07096352
#> 100 100  0.86878947 -1.02106470

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001 -0.43398306 -1.656335925
#> 2   002  0.28869306  0.667011087
#> 3   003  0.49918719  0.732258863
#> 4   004  0.49305379  0.427530679
#> 5   005  0.41065434 -0.286664014
#> 6   006 -1.44694462 -2.234269745
#> 7   007 -1.46622946 -2.157601200
#> 8   008 -0.63796653 -0.049130186
#> 9   009 -1.56674554 -0.864247927
#> 10  010 -0.76886321  1.794381457
#> 11  011  0.18191787  1.583604067
#> 12  012  1.43830157 -0.590093677
#> 13  013 -1.67293131 -1.113147321
#> 14  014  1.36902337 -1.515524667
#> 15  015  0.15408926  0.790492178
#> 16  016 -0.71273269  1.149202448
#> 17  017  0.81572043 -0.686891316
#> 18  018 -0.21827348 -0.304934395
#> 19  019 -0.94477533 -0.510993687
#> 20  020 -0.08463174 -0.744317458
#> 21  021 -0.09806023 -0.946626553
#> 22  022 -0.87552214 -1.495371337
#> 23  023  0.96452400  0.689158363
#> 24  024  0.68964485 -0.601232788
#> 25  025 -2.13141237 -0.487640107
#> 26  026  1.11580155  0.395893015
#> 27  027 -0.85365046 -1.112366304
#> 28  028 -1.04632276 -0.208718827
#> 29  029 -0.13691568 -0.485631226
#> 30  030 -2.30478716 -1.145187400
#> 31  031  0.04156337  1.004692863
#> 32  032  0.39046831 -0.005069972
#> 33  033  0.34296401 -0.894227257
#> 34  034  0.48843564  1.110197516
#> 35  035  0.30029081  0.199561999
#> 36  036 -0.28954728  0.395492766
#> 37  037  2.14283778  0.083459567
#> 38  038 -0.17421276 -0.079157222
#> 39  039 -2.26285284 -2.154858221
#> 40  040 -1.23045166 -1.694844663
#> 41  041 -0.74363486 -1.883506641
#> 42  042  0.69363346  0.781005116
#> 43  043  2.78544058 -0.370022465
#> 44  044 -0.91389394 -1.350397940
#> 45  045 -0.17283946 -0.821984346
#> 46  046 -1.44551296 -1.041172590
#> 47  047 -0.26108382 -1.451674966
#> 48  048  0.35663124  0.809778257
#> 49  049 -0.19868495  0.166952902
#> 50  050 -0.51607733 -2.371482452
#> 51  051 -0.25551334 -0.322612693
#> 52  052  0.01360499 -0.390736150
#> 53  053  0.33973748 -1.814410859
#> 54  054  1.17673523 -1.742228394
#> 55  055 -1.11849119 -0.374630083
#> 56  056 -0.59301908  0.220897833
#> 57  057  1.24923856 -1.538651704
#> 58  058 -0.56755171 -1.789466408
#> 59  059 -0.55500861 -0.168356386
#> 60  060 -1.34882917 -0.734533413
#> 61  061  0.48010685  0.567040491
#> 62  062  0.13477133 -0.992645282
#> 63  063  0.27785195  0.078223294
#> 64  064  0.53256400  0.734859709
#> 65  065  0.88634475 -0.069991629
#> 66  066  0.08756502 -0.484219517
#> 67  067  1.58463540  1.747837988
#> 68  068  0.63241556 -0.461228112
#> 69  069  1.33642379  1.553740465
#> 70  070 -0.01773689 -2.457609994
#> 71  071  0.36808526  0.127613170
#> 72  072 -0.53761818 -2.275654493
#> 73  073  1.50345510  0.907585145
#> 74  074 -0.24548550 -0.235035214
#> 75  075  1.19455486 -0.090799276
#> 76  076  2.73900863  1.418537560
#> 77  077  0.72141718 -0.343509922
#> 78  078  0.20504911 -0.260641938
#> 79  079 -0.52192732  0.782860292
#> 80  080  0.84915748 -0.172727784
#> 81  081  0.18112264  0.398114285
#> 82  082 -0.52829345 -0.818853258
#> 83  083 -0.79504899 -1.159269565
#> 84  084 -0.64063556 -0.821578039
#> 85  085  1.20646946  0.086818286
#> 86  086  1.09240314  0.378522065
#> 87  087  1.30908274  1.390180720
#> 88  088  0.94448263 -0.506927714
#> 89  089  2.66891700  3.260143142
#> 90  090  0.07732159 -0.427814656
#> 91  091  0.26432249 -0.256443243
#> 92  092  0.84329694  1.647323797
#> 93  093 -0.62689558 -0.601742726
#> 94  094  1.17215709 -0.537024377
#> 95  095 -0.55902534 -0.408336309
#> 96  096 -1.34855920 -0.244514626
#> 97  097  0.77173553 -0.539225592
#> 98  098  0.18649979 -0.183086527
#> 99  099 -0.50416669 -0.413796326
#> 100 100 -0.13317119  1.590633896