Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4828773
#> Y_2 0.4828773 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.4575280  0.5457230
#> 2  1.2751411  1.3874165
#> 3  0.5033692 -0.2015925
#> 4 -0.3371126 -0.3338424
#> 5  0.1577307 -0.1961766

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001 -0.75144215  0.475359934
#> 2   002  0.01880462  0.190373666
#> 3   003 -0.40916135 -0.364304488
#> 4   004  1.21922675  2.286043143
#> 5   005 -0.64678515 -1.391179783
#> 6   006 -0.75399594 -0.903682597
#> 7   007  1.22065694  2.295466406
#> 8   008  1.49332277  0.050213131
#> 9   009 -0.90835865  0.646822648
#> 10  010 -0.47624154 -0.140992043
#> 11  011 -0.58139976 -0.372678670
#> 12  012  0.02575993 -0.628789094
#> 13  013  0.39967924 -0.897193178
#> 14  014 -0.12126566 -0.242963789
#> 15  015 -1.30134480 -1.806651991
#> 16  016 -0.13592968 -0.855491992
#> 17  017 -0.10622904 -1.091899299
#> 18  018 -1.12931804 -2.375618780
#> 19  019  0.91369832  0.843565258
#> 20  020 -0.33619243 -0.036189392
#> 21  021 -0.52133813 -0.255885520
#> 22  022 -1.12988045  0.103029058
#> 23  023  0.58559942  0.393148466
#> 24  024 -1.70482582 -0.985226018
#> 25  025 -0.58646687 -0.190419679
#> 26  026  0.71344474  0.327586106
#> 27  027 -0.10666108 -0.381820639
#> 28  028 -0.18373844 -0.635731876
#> 29  029 -0.34748952 -0.258453682
#> 30  030 -1.67432393 -0.994424848
#> 31  031 -0.63832312 -0.754043439
#> 32  032 -0.69571033 -1.713095431
#> 33  033 -0.53648863  0.689872249
#> 34  034  0.71477772  0.382644155
#> 35  035 -2.45506315 -1.351217123
#> 36  036 -1.95415389 -0.395728126
#> 37  037  0.60735604  0.071066682
#> 38  038  0.42873844 -0.126060393
#> 39  039  0.06795340 -1.788630907
#> 40  040 -1.77670219 -1.001489884
#> 41  041  1.72061243  0.392072398
#> 42  042  0.18212724  0.432193914
#> 43  043 -0.17166371 -0.559432481
#> 44  044 -0.50639050  0.367112924
#> 45  045 -0.88703796  0.330058683
#> 46  046 -0.98028213  0.728094539
#> 47  047  0.44065476 -0.222406665
#> 48  048 -0.80639914 -2.153555177
#> 49  049 -0.44091822 -0.845535475
#> 50  050  0.84194787  0.893266484
#> 51  051 -0.47984301  0.060212786
#> 52  052 -0.46122318 -0.197909808
#> 53  053  0.89383924  0.626627015
#> 54  054  0.08485794 -0.609623481
#> 55  055 -0.18319780 -0.186654019
#> 56  056 -1.54729977  0.488600468
#> 57  057  1.69936091 -0.605869282
#> 58  058  1.07059803  1.005049015
#> 59  059 -1.09447782 -0.927972232
#> 60  060  1.11249747 -1.035587644
#> 61  061 -1.53261099  0.218542386
#> 62  062 -0.45249284  0.696299505
#> 63  063  0.14710814  0.234629559
#> 64  064 -0.16132749 -0.007036567
#> 65  065  1.02896892  1.984663390
#> 66  066  0.44683758  0.956020334
#> 67  067 -1.20327014 -0.675351314
#> 68  068  2.03060680 -0.193401682
#> 69  069  0.13427999  0.113142815
#> 70  070 -0.94184265 -2.147211811
#> 71  071  0.05988990 -1.212890764
#> 72  072 -0.14103310  1.827787024
#> 73  073 -0.52943021  0.384857390
#> 74  074 -0.56208013  0.941079690
#> 75  075  0.54379238  0.812434819
#> 76  076 -0.16610798 -1.213206830
#> 77  077 -0.31232832 -0.436186263
#> 78  078  1.51547113  0.598504900
#> 79  079  0.02958696 -0.252938262
#> 80  080 -0.84890488 -0.299596175
#> 81  081  1.28683600  0.250892246
#> 82  082  1.87417847  0.742917721
#> 83  083  0.61065865  0.871656901
#> 84  084  0.33662154 -0.870469406
#> 85  085  0.13282925 -1.407624259
#> 86  086  0.41020968  1.326530046
#> 87  087 -0.64724825 -1.271925777
#> 88  088  0.42933615 -0.464321664
#> 89  089  0.37540364 -0.696314737
#> 90  090  0.22709555  0.558788499
#> 91  091 -0.05760396 -1.175328513
#> 92  092 -1.03505559 -1.034671830
#> 93  093  0.23118342  0.933557554
#> 94  094  1.56621602  0.721994075
#> 95  095  0.69523985  0.321247338
#> 96  096  0.66977159  0.498775794
#> 97  097  0.44856502 -0.523826998
#> 98  098  1.10402890  1.555954512
#> 99  099 -0.38076479  0.049883327
#> 100 100 -1.30758346 -0.794794046

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID           Y.1         Y.2
#> 1   001 -0.8998934027 -0.65796439
#> 2   002 -0.7233621591 -1.20834131
#> 3   003 -0.3247154542 -1.02659486
#> 4   004 -1.0073657080 -0.71036360
#> 5   005  0.5900067093 -0.53221917
#> 6   006  0.4647820783  1.38294118
#> 7   007 -0.8489116803  0.76420691
#> 8   008  0.0270122460 -0.19687214
#> 9   009 -0.4993051667  0.11372583
#> 10  010  0.4369085976 -0.14252908
#> 11  011 -0.1508178687  1.44744849
#> 12  012 -0.7795051096 -1.23100172
#> 13  013  0.9724410320  0.27728062
#> 14  014  0.4083107736 -0.68182888
#> 15  015  1.4095468613 -0.01340457
#> 16  016  0.2140509702  0.08543756
#> 17  017  1.3524732926  1.64351919
#> 18  018  0.1791653798 -1.12446744
#> 19  019  0.7614395644  0.20781960
#> 20  020  0.0831065362  0.64766710
#> 21  021 -1.9128035778  0.77277505
#> 22  022 -0.1477875415 -0.49455840
#> 23  023 -0.9219078180 -0.08123153
#> 24  024  0.4422321089  1.84980853
#> 25  025 -0.5983466597  0.21767369
#> 26  026  0.3931426527  1.15612689
#> 27  027  0.1581150013 -1.42811810
#> 28  028 -2.4658129616  0.20315987
#> 29  029  1.4215617580  0.19070732
#> 30  030  0.8415170532 -0.42223256
#> 31  031  1.0245795939  0.50352980
#> 32  032  1.5902961221  1.36643491
#> 33  033  1.0208500674  1.50878711
#> 34  034  1.0806381410 -0.08569266
#> 35  035 -0.7318092949  0.44801061
#> 36  036  1.0207293284 -0.14722884
#> 37  037  0.3131333593  1.09945055
#> 38  038  0.6601751843  0.77760005
#> 39  039 -0.5834520099 -1.29463270
#> 40  040 -0.6910919234 -2.11371678
#> 41  041 -0.1984213436 -0.07226154
#> 42  042 -0.3750520252 -1.73064738
#> 43  043  0.9336308646 -0.58332018
#> 44  044 -0.0914187250  0.34516473
#> 45  045  0.1192752677  1.44585476
#> 46  046  1.2099955378  1.18207862
#> 47  047  0.2199596481 -0.51026785
#> 48  048  0.4920841566 -0.92925877
#> 49  049 -1.4809030177 -0.62875625
#> 50  050  0.6402730262  0.11254297
#> 51  051 -1.1652472627 -0.81756991
#> 52  052 -1.5800597934 -0.65828579
#> 53  053  0.7437160431  2.17263492
#> 54  054  1.0929412813 -1.11785416
#> 55  055 -0.5942476925  0.65054845
#> 56  056  0.1070815900  0.53750521
#> 57  057 -1.7435512209 -1.53690883
#> 58  058  1.2992865977  0.79559350
#> 59  059 -0.1215713015 -1.21447382
#> 60  060 -1.0200657653  1.86796575
#> 61  061  0.2077100359  0.79396033
#> 62  062  1.1183291830  0.16017310
#> 63  063 -0.4594973723 -1.66574072
#> 64  064 -0.8395106673 -0.47842512
#> 65  065  1.0259280211 -0.27602831
#> 66  066 -1.7642593397 -1.56246548
#> 67  067  0.1401438821  1.35222860
#> 68  068 -1.8563205815 -1.11949851
#> 69  069  0.5291413988  0.60334092
#> 70  070  1.5940386773  0.73076369
#> 71  071  1.6623073243  1.97417527
#> 72  072  0.0002825186  0.56756494
#> 73  073  0.0368512698 -0.35314655
#> 74  074 -1.5019179075  1.03138300
#> 75  075  1.7437951733  0.79272326
#> 76  076  0.2827310101 -1.21539009
#> 77  077  1.2924063168  2.12064486
#> 78  078 -0.5600148477 -0.48811368
#> 79  079  1.2765476574 -0.57000107
#> 80  080  0.2743819773  0.12548961
#> 81  081 -1.2132570619 -1.14336221
#> 82  082  0.6442220376  0.08020059
#> 83  083 -0.1614625408  0.56636947
#> 84  084 -0.1017001624  0.75783284
#> 85  085  0.6199780129  0.67506737
#> 86  086  0.8939900327 -0.62540253
#> 87  087  1.2455550707 -0.10477445
#> 88  088  1.4794105761  1.58330836
#> 89  089 -0.1875903124 -0.20857596
#> 90  090 -0.0413969952  1.71190193
#> 91  091 -0.1745442274 -0.22709594
#> 92  092  1.3771698857  0.64361496
#> 93  093 -1.4448952602 -1.20297890
#> 94  094 -1.2214051911 -1.17185442
#> 95  095 -0.2246288111  0.05382582
#> 96  096 -0.4794015168 -0.96826082
#> 97  097 -0.5194080993  0.51563968
#> 98  098  0.3046529139  0.68024165
#> 99  099  1.1354566643  0.05441290
#> 100 100 -0.6261217899 -0.81587556