Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5107855
#> Y_2 0.5107855 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1 -0.8558328  0.1814364
#> 2  0.3488906  1.9755647
#> 3 -1.5378832  0.5632476
#> 4 -1.0019264 -0.5035387
#> 5 -0.7818857  0.3857109

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  1.058293287  1.428763942
#> 2   002 -0.392587893  1.037042523
#> 3   003 -1.384940251 -0.982301304
#> 4   004  0.547700749  0.353260000
#> 5   005 -0.854495363 -0.750304540
#> 6   006  1.949450415  1.146843159
#> 7   007  0.445535188 -0.884670667
#> 8   008 -0.328136180 -0.872415896
#> 9   009 -0.254831645  2.174738499
#> 10  010  1.120251567  1.698017970
#> 11  011  1.295704158  1.251223804
#> 12  012  1.201612758 -0.335922888
#> 13  013  0.051362655 -0.348344279
#> 14  014  1.258798474  1.284399395
#> 15  015 -0.717475598  0.044081893
#> 16  016 -0.021344130 -1.618883692
#> 17  017  0.555740145  0.435542024
#> 18  018  1.964060234  0.807575281
#> 19  019 -2.150166316 -1.403406280
#> 20  020  0.279664751  0.176336547
#> 21  021 -1.124730480 -1.105047052
#> 22  022  0.702032739  0.560125086
#> 23  023 -1.134403000 -0.368627164
#> 24  024 -0.403554461 -1.639358021
#> 25  025  1.400201006  0.681478878
#> 26  026  1.244999232  0.698426746
#> 27  027  1.265496283 -1.119579669
#> 28  028  0.217746565 -0.744725822
#> 29  029  0.202427214  0.411421887
#> 30  030 -0.304439379 -0.781500688
#> 31  031  0.416791418 -0.255365944
#> 32  032  0.038621406  1.077339992
#> 33  033  0.774960773  0.336101691
#> 34  034 -1.454808721 -0.500206854
#> 35  035 -1.336119194 -1.213435530
#> 36  036 -0.082020468  1.142573566
#> 37  037 -0.866432598  0.665061687
#> 38  038 -0.191634308 -0.921205547
#> 39  039 -1.096854856 -1.877378237
#> 40  040  0.127225571 -0.194823061
#> 41  041 -0.335155848 -0.299323488
#> 42  042  0.896208181  0.674966560
#> 43  043  1.012781342 -0.819977839
#> 44  044  1.442866260  0.333443633
#> 45  045 -0.012772049 -0.504526092
#> 46  046  0.742550307  1.091954560
#> 47  047  0.091240441  1.776114740
#> 48  048 -0.571883390 -0.948784534
#> 49  049  0.492357453 -0.307807656
#> 50  050 -0.192793281  0.059347356
#> 51  051  0.076625212 -0.371261812
#> 52  052  0.844618774  1.911825487
#> 53  053  0.470049224  0.564159358
#> 54  054 -0.039619367 -0.883318159
#> 55  055  0.222405809 -0.392540706
#> 56  056 -0.098687633 -0.030323019
#> 57  057 -0.780230524 -1.601992992
#> 58  058  1.531218093  1.151668732
#> 59  059 -0.326869547  0.249845578
#> 60  060  0.593313676 -0.467195626
#> 61  061 -0.438498117 -0.651703154
#> 62  062 -0.241609650 -0.802042679
#> 63  063 -0.331623793  0.330521585
#> 64  064  0.823741677 -0.602948395
#> 65  065 -0.594736976 -1.248901408
#> 66  066 -0.111838068  0.199504327
#> 67  067  0.437101727 -0.570422876
#> 68  068 -0.973393212 -0.414821304
#> 69  069  1.015388857  0.392086834
#> 70  070  0.283245173 -0.177093750
#> 71  071  0.006665661  1.245261393
#> 72  072 -0.159177987 -0.789651571
#> 73  073 -1.578998468 -0.679242107
#> 74  074 -0.762745094  0.358075040
#> 75  075 -1.214437117  1.129692589
#> 76  076 -0.281461629  0.839454775
#> 77  077 -2.118825171 -0.956486835
#> 78  078  0.650588743  0.033311978
#> 79  079  0.193465050  0.696393226
#> 80  080 -1.266053961 -0.071485614
#> 81  081 -0.402414138 -0.983986630
#> 82  082 -1.506918811 -1.251166923
#> 83  083  0.937299877 -0.329353343
#> 84  084  0.760407367  0.510351570
#> 85  085 -0.516863739  0.019889327
#> 86  086  1.895986065  1.703111449
#> 87  087 -0.442417335  1.039765717
#> 88  088 -0.514690276 -0.046294003
#> 89  089 -0.136967337 -0.983614735
#> 90  090  0.320427359  0.547509808
#> 91  091 -0.211970995 -1.097541078
#> 92  092 -0.592626337  0.006735459
#> 93  093 -0.578729924  1.693619280
#> 94  094  0.664053124  0.274986729
#> 95  095  1.042348790  0.684142251
#> 96  096 -1.211984767 -0.267785952
#> 97  097 -0.848955808 -0.708621018
#> 98  098 -0.856592159  0.410075715
#> 99  099 -0.857393684  0.515640245
#> 100 100 -0.633041285  0.137016507

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -1.91966139 -0.05406762
#> 2   002  0.13431391 -1.49145522
#> 3   003 -1.10973204 -0.61793253
#> 4   004 -1.12631608 -1.27208661
#> 5   005  1.46276154  1.27278303
#> 6   006  1.59985246 -0.36058889
#> 7   007 -0.93021485 -1.00083533
#> 8   008  0.02682128 -0.81222996
#> 9   009 -0.57824412 -0.10113093
#> 10  010  1.31744425 -0.66010000
#> 11  011 -0.28646801 -0.11914969
#> 12  012 -0.30354568 -0.75931551
#> 13  013  0.49429894 -0.80893427
#> 14  014 -0.61295500 -0.55811734
#> 15  015 -1.57498090 -0.82635988
#> 16  016 -1.48317004 -0.28247627
#> 17  017  0.69869398  0.82162512
#> 18  018 -0.68333610  0.82400439
#> 19  019 -1.05372981  0.37336242
#> 20  020  0.70770515  0.18591813
#> 21  021  2.00819682  0.65680820
#> 22  022  0.79673459  0.06534497
#> 23  023  0.11445686  0.19113650
#> 24  024  0.24412906  1.17772427
#> 25  025  3.13826177  2.49009714
#> 26  026 -0.43702945 -1.06485780
#> 27  027  0.53863519  1.07149473
#> 28  028  0.91383848 -1.29424982
#> 29  029 -1.01808367 -1.33802221
#> 30  030  1.00427632  1.01278716
#> 31  031 -0.40795405  0.09821329
#> 32  032  1.10317674  0.92569218
#> 33  033  0.04717119  0.99967166
#> 34  034  0.03772433 -1.26637415
#> 35  035  0.29700757  0.41605533
#> 36  036  0.67644407  2.08991714
#> 37  037  1.65339433  0.66237479
#> 38  038 -0.04361897 -0.25502476
#> 39  039  0.66992275  0.24648430
#> 40  040 -1.45428283 -0.12255854
#> 41  041  0.67166262  0.65084816
#> 42  042  0.02492527 -0.49725744
#> 43  043  1.84494376  1.58074517
#> 44  044  1.53290112  1.30372643
#> 45  045  0.57510786 -0.38475739
#> 46  046  1.07679955 -0.54224720
#> 47  047 -1.38876057 -0.91672522
#> 48  048  1.30231860  0.73016823
#> 49  049 -1.64998484 -1.09628027
#> 50  050  0.95720498  0.41759079
#> 51  051  0.21473435 -0.16813444
#> 52  052 -0.27957605  1.19962845
#> 53  053 -1.23026423 -0.85992326
#> 54  054  0.12291419 -0.72054843
#> 55  055 -1.05942716 -0.46590570
#> 56  056 -1.16573914 -1.24027635
#> 57  057  0.77481542 -0.14233621
#> 58  058 -0.73867081 -0.69853715
#> 59  059 -0.62784334 -0.48558647
#> 60  060 -0.31589113 -0.65195587
#> 61  061 -0.75481436 -0.99178352
#> 62  062  0.22719556  0.31450678
#> 63  063  0.95640165 -0.55204157
#> 64  064  1.15096212  0.58937221
#> 65  065 -0.87292864 -0.86726376
#> 66  066  1.76586565  0.58081617
#> 67  067 -0.46963828  0.40146760
#> 68  068  1.18610322  0.93122196
#> 69  069  1.63292524  0.94713346
#> 70  070  1.07227884 -0.89775962
#> 71  071 -1.35632284 -0.51940321
#> 72  072  1.16032700  0.55491781
#> 73  073 -1.88093858 -2.09403999
#> 74  074 -0.45826701 -0.11455555
#> 75  075 -0.24070879  0.56030057
#> 76  076 -0.06535460 -1.54919494
#> 77  077 -0.22667236  0.28913869
#> 78  078 -1.41322748 -0.68142811
#> 79  079 -0.78666620  0.52835661
#> 80  080  0.37967716 -0.04879788
#> 81  081 -0.52482407 -1.71210995
#> 82  082  0.38545227  1.76802920
#> 83  083 -0.55840084  0.41214579
#> 84  084  0.01450317 -2.12228463
#> 85  085 -0.17869032 -0.04996942
#> 86  086  1.59854385  0.58030065
#> 87  087 -0.69773252 -0.87838282
#> 88  088 -0.08812405 -1.10861664
#> 89  089 -0.43677254 -0.79747324
#> 90  090 -0.19354416 -0.11857938
#> 91  091 -0.54999761  0.03088488
#> 92  092  1.12222675  0.43676171
#> 93  093 -0.16651254 -1.42800747
#> 94  094 -1.84292049 -1.12705659
#> 95  095 -0.50693873  0.52017852
#> 96  096  0.29589945  2.42418664
#> 97  097 -1.09885675 -0.12137017
#> 98  098  1.05765975  0.92291444
#> 99  099 -1.76409716 -1.42589237
#> 100 100 -1.24391769  0.32802033