Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5527416
#> Y_2 0.5527416 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>           Y_1        Y_2
#> 1 -1.34269693 -0.4731693
#> 2 -1.11108912  0.2475808
#> 3 -0.08250095  0.4017281
#> 4 -2.12753495 -0.7268956
#> 5 -1.92171330 -1.2825495

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001 -0.324120624 -0.434377979
#> 2   002  0.593905541  0.672128725
#> 3   003 -0.180599807  0.040706478
#> 4   004 -0.371252228  0.300329736
#> 5   005  0.861856509  1.487904843
#> 6   006  1.114288006 -0.343441318
#> 7   007 -0.596574856 -1.081059066
#> 8   008  0.195096495  0.716686051
#> 9   009 -0.021880247 -0.348108912
#> 10  010  0.542858954  0.042250900
#> 11  011 -1.677823523 -2.106190593
#> 12  012  1.816263013  0.153516477
#> 13  013  1.842880133 -0.169777526
#> 14  014 -0.470060622 -0.030110554
#> 15  015  0.258680508 -0.198570305
#> 16  016  0.261589810 -0.794252861
#> 17  017  0.140872629 -1.879669009
#> 18  018 -1.915946120 -1.439741100
#> 19  019 -0.758251214  0.891608771
#> 20  020 -2.872645940 -1.515802871
#> 21  021  1.909051418  1.292310421
#> 22  022 -0.026263103 -0.642075346
#> 23  023 -1.134344939  0.069087215
#> 24  024 -0.147222625 -0.814675955
#> 25  025 -0.763115507 -1.392878883
#> 26  026  0.947692019 -0.102357841
#> 27  027  0.121044708  0.787451955
#> 28  028  1.446537469  0.059253161
#> 29  029 -0.128796478 -0.899054070
#> 30  030  2.151805534  1.493555382
#> 31  031 -1.073618316 -0.724688270
#> 32  032  1.694115066  0.549475684
#> 33  033 -0.352518340 -0.237107222
#> 34  034 -2.033559688 -2.170853393
#> 35  035  0.111841949  1.413773260
#> 36  036 -0.338552369  0.707465779
#> 37  037 -1.181620139 -2.298524993
#> 38  038 -0.468930843 -1.104519751
#> 39  039 -0.989173987 -1.142918185
#> 40  040  1.590013889  0.906183717
#> 41  041  0.039892081 -0.220514314
#> 42  042  0.627711285  0.626431969
#> 43  043 -0.587129513 -0.749510595
#> 44  044 -0.387436264 -0.457643717
#> 45  045  1.323793433  1.098038869
#> 46  046  0.396275417  0.723587461
#> 47  047 -0.153870132 -0.035532640
#> 48  048  1.190431704  2.067577652
#> 49  049  0.590513775 -0.032952300
#> 50  050  0.257042133  1.302095563
#> 51  051  1.597932357 -0.324935483
#> 52  052 -0.126312319 -0.416486805
#> 53  053 -0.421973142 -0.995784764
#> 54  054 -0.255993588 -0.600429094
#> 55  055 -0.703211224  0.140547225
#> 56  056 -0.218319841 -0.585490402
#> 57  057 -0.259164595  0.949688760
#> 58  058  0.889835172 -1.352229595
#> 59  059  0.002170216  0.265392203
#> 60  060 -0.284579665  0.963770380
#> 61  061 -0.877543864  0.095559082
#> 62  062 -0.991639262 -1.145803452
#> 63  063 -1.578943268 -0.630982280
#> 64  064 -2.078488892 -0.249615298
#> 65  065  2.514946397  1.583173299
#> 66  066 -1.574404660 -0.634392125
#> 67  067 -0.046032267  0.100901172
#> 68  068 -1.771050634 -0.159950758
#> 69  069 -0.226656793  0.998935220
#> 70  070 -0.590080145 -1.620072335
#> 71  071  0.400066564 -0.493585056
#> 72  072 -0.117981160  1.181634996
#> 73  073 -0.206507471 -0.610819006
#> 74  074  0.520345717  0.633369940
#> 75  075  1.474397830  0.905145967
#> 76  076 -1.774808693 -0.666274546
#> 77  077  0.370797540 -0.599947637
#> 78  078 -0.114318588  1.039446221
#> 79  079  0.858067810 -0.259007368
#> 80  080 -0.444549531  0.529362893
#> 81  081  1.157746412  1.417467713
#> 82  082  0.085406857  0.897683279
#> 83  083  0.535434620 -1.153121641
#> 84  084  0.111383702 -1.748595542
#> 85  085 -1.635193091 -1.225276505
#> 86  086 -1.256939019 -0.100277589
#> 87  087 -0.561417888 -0.055282314
#> 88  088 -0.064598316  0.002497632
#> 89  089 -0.608102172  0.166791985
#> 90  090  0.068148625  0.921429750
#> 91  091  0.618923289  0.606414153
#> 92  092 -0.565283395  0.535603354
#> 93  093  0.757862866  0.482910793
#> 94  094  0.699084108  0.198238896
#> 95  095  0.541079523  1.382725869
#> 96  096 -0.215520134 -1.766391172
#> 97  097 -1.048157163 -1.313168488
#> 98  098  0.646875201  0.165161191
#> 99  099 -0.080875281  0.795296039
#> 100 100 -1.028884521 -0.528296157

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -0.274024831 -1.53499057
#> 2   002 -0.116693625  0.49820483
#> 3   003  0.964225877 -0.84477396
#> 4   004 -1.053919444 -0.12375121
#> 5   005  0.997938112  0.42109254
#> 6   006 -1.199154188 -1.13941132
#> 7   007  0.720128299  1.53993983
#> 8   008  0.875166365  0.62934091
#> 9   009  0.548984951 -0.62613259
#> 10  010  1.308606572  0.12677341
#> 11  011  0.896868590  0.68666447
#> 12  012  0.227741520  0.59408921
#> 13  013 -1.350651678 -1.83341736
#> 14  014  0.377538825  0.75810618
#> 15  015  1.262282428  0.61188447
#> 16  016 -0.399857361 -1.76243080
#> 17  017 -1.037094706 -0.26614898
#> 18  018  0.686045245  0.17566050
#> 19  019  0.676705043  0.31069406
#> 20  020 -0.252903604 -0.99946587
#> 21  021  0.456331568  1.74964260
#> 22  022 -0.071119614 -0.26608052
#> 23  023 -0.383257368  1.32446335
#> 24  024  0.772706578  0.53810897
#> 25  025  0.085205915  0.79690957
#> 26  026  0.649470402  0.16781620
#> 27  027 -1.954339274 -0.68732775
#> 28  028  0.964959632  1.03713146
#> 29  029  0.453194822 -0.53819470
#> 30  030 -0.531103514 -0.66588127
#> 31  031 -0.641394831  0.22577236
#> 32  032 -0.374840142 -0.76041096
#> 33  033  0.199843510  1.67207054
#> 34  034 -0.206136227 -0.24192048
#> 35  035 -1.047791317 -0.12954389
#> 36  036 -0.119050632  0.91610446
#> 37  037  1.780910207  0.50587359
#> 38  038 -0.150987739 -0.04845835
#> 39  039 -0.225454104  0.34865867
#> 40  040 -1.891955138 -0.50621582
#> 41  041  0.980342268  0.65258149
#> 42  042  0.326466627  0.26874905
#> 43  043  0.092150808 -1.57496142
#> 44  044 -0.350613559 -1.57089924
#> 45  045  0.273676985 -0.66995575
#> 46  046  0.382031478 -0.23884382
#> 47  047  0.890228866 -0.03323397
#> 48  048 -0.826720647  0.20322677
#> 49  049 -0.846631234 -1.00867947
#> 50  050  0.932117358 -0.35664871
#> 51  051  0.033125900 -0.90213576
#> 52  052  0.403415957 -1.73528467
#> 53  053  0.534519730  0.12010184
#> 54  054  1.580576193  0.85952404
#> 55  055 -0.011561376 -0.10139674
#> 56  056 -2.062568052 -0.77826102
#> 57  057  0.921255382 -0.54959479
#> 58  058  0.833432274  1.05509054
#> 59  059 -0.057600139  0.08449027
#> 60  060  0.855782857 -0.65771791
#> 61  061  1.513158582  0.19510595
#> 62  062 -0.222967963  0.31893392
#> 63  063  0.642988664  1.09193595
#> 64  064  0.609493658 -0.40428396
#> 65  065  1.184535722  1.49443625
#> 66  066 -0.516084183 -0.48970921
#> 67  067  0.317986952 -1.49666148
#> 68  068 -0.111748478 -0.23689909
#> 69  069  0.286785003 -0.71423759
#> 70  070 -0.283099471 -0.43746900
#> 71  071 -1.259593897  0.25681937
#> 72  072  0.959822488 -0.18746182
#> 73  073 -1.360575439 -0.58273536
#> 74  074  1.514468761  0.20758283
#> 75  075  1.738807570  0.78209260
#> 76  076 -0.872418433 -0.41590498
#> 77  077  2.383512549  0.09606180
#> 78  078 -1.057593717 -0.83150536
#> 79  079  0.180462405  0.66605148
#> 80  080  2.078283025  1.13678995
#> 81  081  1.459820608  0.42758148
#> 82  082  0.731687221 -0.09259936
#> 83  083  0.864963506  1.19640929
#> 84  084 -0.004143486  0.26135563
#> 85  085  0.778317995  0.31538969
#> 86  086  0.946066168 -0.16709005
#> 87  087 -1.925992017 -0.02908078
#> 88  088  1.117352485  1.37470550
#> 89  089 -0.675420994 -0.54736013
#> 90  090 -0.293749822 -0.03051153
#> 91  091 -0.884664533 -0.23832785
#> 92  092  0.803331262 -1.45080566
#> 93  093  2.464002655  1.59182287
#> 94  094  1.180882859  1.06810251
#> 95  095  1.431112142  0.06043550
#> 96  096 -1.830492142 -0.30009488
#> 97  097  0.613845068 -0.90020397
#> 98  098 -0.715369597 -1.98731219
#> 99  099 -0.970174245 -0.86976202
#> 100 100 -2.746813292 -1.19046118