Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5333122
#> Y_2 0.5333122 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.0408362 -1.24507611
#> 2  0.3627761  0.95677926
#> 3 -0.4250167  0.05723813
#> 4  1.0202021  1.35791175
#> 5 -1.0220971  0.06729044

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1           Y_2
#> 1   001 -0.86472506 -0.2301019768
#> 2   002  0.90838685  0.1559493572
#> 3   003  1.07908801 -0.5682674910
#> 4   004 -0.55733362 -0.4996251150
#> 5   005  2.68537645  0.5294394948
#> 6   006  0.29848099 -1.8259960263
#> 7   007  0.46800402 -1.0029831170
#> 8   008  0.22232199  1.3623136630
#> 9   009 -0.47831544  0.1492286475
#> 10  010 -1.07122662 -1.6215268867
#> 11  011  0.62388468  0.7428784246
#> 12  012  0.29624927 -0.5254626339
#> 13  013 -0.39979066 -0.3849633149
#> 14  014 -2.04680564  0.0193681311
#> 15  015  1.76490154  2.7001123833
#> 16  016 -1.00728267 -1.6568128873
#> 17  017 -1.36840740 -1.1389357793
#> 18  018 -0.90345793  0.0040544173
#> 19  019  1.07800942 -0.7821659294
#> 20  020 -2.01723952 -0.2725894548
#> 21  021 -1.71012137  0.0544955015
#> 22  022 -0.22520058  0.8164452103
#> 23  023 -0.48809096 -0.2131445358
#> 24  024 -0.78625748 -0.4177577252
#> 25  025 -0.77117334 -1.4716811137
#> 26  026 -0.08846867  0.9474492766
#> 27  027  1.08404858  1.2123696952
#> 28  028  0.83567135  2.0852321152
#> 29  029 -1.47700621 -0.3679596250
#> 30  030 -0.57315400 -0.7891580329
#> 31  031 -0.25461436 -0.0432294187
#> 32  032  0.64278367 -0.8196737465
#> 33  033 -0.90688884 -1.2162256587
#> 34  034 -0.15117268 -1.0396225060
#> 35  035  0.13565801  0.1897959652
#> 36  036 -0.29993247  1.1636279214
#> 37  037  0.62254127  0.2946742073
#> 38  038 -0.65416644 -0.0000341897
#> 39  039 -0.45213972  0.1471751993
#> 40  040  0.46579269  0.2443134592
#> 41  041  0.79097660 -0.0098817980
#> 42  042 -1.73144780  0.3668902375
#> 43  043  0.87967580  2.1669892743
#> 44  044 -1.49512253 -0.9474981982
#> 45  045  1.52956770  1.1034216226
#> 46  046  2.41014432  0.6780844133
#> 47  047 -1.36138927 -0.2796537508
#> 48  048 -0.07490185 -0.3531321244
#> 49  049  0.59781846  0.8256910920
#> 50  050  1.77221302  1.8813921418
#> 51  051 -0.46792643  0.6865013465
#> 52  052  0.45995020  1.2896041597
#> 53  053  0.63642854 -0.2415248559
#> 54  054 -1.05882209 -1.1466532362
#> 55  055 -0.61909645  0.1580098055
#> 56  056  0.35644715 -0.7218072020
#> 57  057 -0.89666707 -0.3378036007
#> 58  058 -0.22910991 -0.4517199452
#> 59  059 -0.33124115 -1.0590267773
#> 60  060 -1.49483849 -0.3484212443
#> 61  061  1.28513879 -0.5489998832
#> 62  062 -0.93760877 -0.6744482741
#> 63  063  0.18686285 -1.0081115098
#> 64  064  0.58788976 -1.7037925412
#> 65  065  0.10164082 -1.3574088667
#> 66  066 -0.23765239 -0.3647734286
#> 67  067 -0.12207769 -0.3051871542
#> 68  068  0.90903305  0.5150680216
#> 69  069 -2.67046668 -2.1450235327
#> 70  070 -0.40182488  0.8842638321
#> 71  071  0.37496344  0.1245435699
#> 72  072  0.41535679  0.1448678398
#> 73  073  0.49942042 -0.9344162325
#> 74  074 -1.76460503 -0.9458994102
#> 75  075  1.97824838  1.8630332239
#> 76  076  1.35432419  0.9968049435
#> 77  077 -1.21668386 -1.4728380532
#> 78  078  0.08606225  0.8103148144
#> 79  079  0.55232537  1.1528454819
#> 80  080 -0.58008409 -0.9236037378
#> 81  081 -0.17347945 -0.2512612606
#> 82  082 -0.01754069  0.2749853196
#> 83  083  0.61331645  0.7799011131
#> 84  084  2.41147465  0.6933475999
#> 85  085 -0.99271551 -1.1856607924
#> 86  086 -1.18839579 -0.5330364159
#> 87  087 -1.96594510 -1.6261844193
#> 88  088 -0.34707501 -0.0836724318
#> 89  089 -0.17177745 -0.0339760734
#> 90  090 -1.93910987 -1.3431805515
#> 91  091  0.18316557  0.2766435247
#> 92  092  0.25056663 -0.4667182251
#> 93  093 -0.84105370  0.1516682842
#> 94  094  1.55913635  0.9504365659
#> 95  095 -0.70059910 -0.9547617128
#> 96  096 -1.62077874 -2.3924703226
#> 97  097  0.13700046 -0.8727295606
#> 98  098 -0.68140107  0.1191511750
#> 99  099 -1.92649841  0.1255136069
#> 100 100 -1.64243375 -2.1044135968

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -1.790048418 -1.11374307
#> 2   002 -0.508698115 -0.54949534
#> 3   003  0.424721752 -0.31542852
#> 4   004  1.524971393  1.20227162
#> 5   005 -0.473065689  0.56963046
#> 6   006 -0.749007914 -1.05755091
#> 7   007 -1.256555901 -0.33807755
#> 8   008 -0.933254092 -0.87390563
#> 9   009 -0.413576941 -0.32853674
#> 10  010  0.324250367 -0.35887287
#> 11  011  0.381141663  0.20913609
#> 12  012 -0.447872026  0.61620303
#> 13  013 -1.015471677 -0.74368454
#> 14  014  0.848947388 -1.13432973
#> 15  015 -1.165589580 -1.20432017
#> 16  016 -0.374732741 -0.61598256
#> 17  017  0.506101623  0.55312097
#> 18  018 -1.531980385 -0.06243625
#> 19  019  0.273521041  0.93027580
#> 20  020  1.475287240  1.42397247
#> 21  021 -1.956045407 -0.22826090
#> 22  022 -0.127121699  0.34936464
#> 23  023 -1.508277492  0.74760547
#> 24  024 -1.228293031  2.76746835
#> 25  025  1.126069405  0.91926573
#> 26  026 -0.021818473 -0.44707732
#> 27  027 -1.087108000 -0.09156246
#> 28  028  2.180712931  1.65445880
#> 29  029  0.173351779 -0.28892384
#> 30  030 -0.557941687 -0.82713845
#> 31  031 -0.207118463  0.16724460
#> 32  032 -0.050683297  0.27776786
#> 33  033  0.873872652  0.65263813
#> 34  034  0.809553572  0.31269543
#> 35  035  0.343721339  0.21132681
#> 36  036  0.202248740 -0.50475096
#> 37  037 -0.549518174 -0.09168265
#> 38  038 -0.842519402 -0.51501668
#> 39  039  2.053011334  1.59278325
#> 40  040  1.691741016  0.05527202
#> 41  041  0.915102209  0.96610557
#> 42  042  0.001424283 -0.53182613
#> 43  043  0.415730939 -0.28671126
#> 44  044 -2.632696695 -3.15465484
#> 45  045  0.967345597  1.00966174
#> 46  046 -0.938818036 -1.30913858
#> 47  047 -0.844797333 -0.99616835
#> 48  048  1.098134817  0.29820376
#> 49  049  1.538233220  0.89860433
#> 50  050  1.447286877  0.78034723
#> 51  051  1.373526619  1.62269567
#> 52  052  0.443870492 -0.69262003
#> 53  053 -0.086997010  0.14415447
#> 54  054  1.680791481  2.06941909
#> 55  055  0.519079190 -0.57116252
#> 56  056  0.138390785 -0.97235447
#> 57  057 -0.897223046 -1.07396597
#> 58  058  1.715777247  1.38217878
#> 59  059 -0.388770611 -0.10304986
#> 60  060  0.184115204 -2.25165048
#> 61  061  0.381946642 -1.09984062
#> 62  062  0.726767654 -1.02928262
#> 63  063 -1.319950135 -2.82946833
#> 64  064 -0.198581898  0.85990968
#> 65  065 -0.025284985 -0.37689191
#> 66  066 -1.027918349 -0.49386530
#> 67  067  0.920645655 -0.47125719
#> 68  068 -0.449418068 -1.53286739
#> 69  069 -1.093727379  0.16151381
#> 70  070  0.557817485 -0.76650603
#> 71  071  1.023612819  0.35846548
#> 72  072  2.537836090  1.15477318
#> 73  073 -0.921428844  0.09695263
#> 74  074  0.406648049  1.10972480
#> 75  075  0.060213618  1.74947237
#> 76  076  0.917546701 -0.66902229
#> 77  077  1.451711062  1.01155068
#> 78  078 -0.744142452  1.44254614
#> 79  079  0.663739563  0.69545649
#> 80  080 -0.591554388 -0.25852393
#> 81  081 -1.223668951  0.03899232
#> 82  082  1.456795193  1.06157141
#> 83  083 -1.310938653 -0.65474833
#> 84  084 -0.272326259  0.51346906
#> 85  085  0.921339001 -1.25718933
#> 86  086 -0.756394292 -2.03964720
#> 87  087  0.791894399  1.64590734
#> 88  088  1.127662120  0.26285391
#> 89  089 -0.729667718 -1.38339566
#> 90  090  1.113767969  1.47370857
#> 91  091  1.117754061 -0.78312300
#> 92  092  0.081398911  0.75291904
#> 93  093 -0.655789858  0.42112758
#> 94  094  0.829483153  1.21196542
#> 95  095 -0.327651555  0.40337548
#> 96  096 -0.509130144  0.58087078
#> 97  097 -1.952135545 -1.00192150
#> 98  098  1.503869512 -0.31582017
#> 99  099 -1.896055548 -1.98841777
#> 100 100 -0.870749136  0.09997734