Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5336612
#> Y_2 0.5336612 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.3513539 -1.22081217
#> 2  0.7668354  1.15199691
#> 3  0.7978139  0.88768255
#> 4  2.0385906  1.27831545
#> 5  0.9954176  0.07824794

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID          Y_1          Y_2
#> 1   001  0.939784395  0.729393249
#> 2   002 -0.831593459  0.031625771
#> 3   003  0.579620217  1.253276858
#> 4   004  0.352186058  1.544327479
#> 5   005  0.987909536  1.454326577
#> 6   006 -1.595941778 -1.187039042
#> 7   007  0.939702074  3.001963313
#> 8   008 -0.633230699 -0.171307037
#> 9   009 -0.523944486 -0.846636836
#> 10  010 -0.591007453 -1.141101538
#> 11  011 -0.605426505 -0.705524156
#> 12  012 -1.427039920 -0.994618884
#> 13  013 -0.314750629  0.013432239
#> 14  014 -1.692517300 -1.005573174
#> 15  015 -1.388966450 -0.511119282
#> 16  016  0.210278518  1.148653806
#> 17  017  0.540073449 -0.077011963
#> 18  018  1.290004356  0.558693770
#> 19  019 -0.012165725  0.765168941
#> 20  020  0.227179582 -0.659789191
#> 21  021  1.203306035  0.659227656
#> 22  022  0.211617761  1.046594915
#> 23  023  1.046806047  0.754392584
#> 24  024  0.645481962 -0.110006337
#> 25  025 -0.281346171  0.264384753
#> 26  026  1.456188770  1.365787360
#> 27  027 -0.367995748  0.254732100
#> 28  028 -0.428026362 -0.718608318
#> 29  029  0.791011770  1.503021020
#> 30  030  0.236956116 -0.405163286
#> 31  031 -0.158025283 -0.878078740
#> 32  032  0.381998837  0.525574969
#> 33  033  0.778781204  0.350893655
#> 34  034 -0.550486132  0.477561392
#> 35  035 -0.164616059 -0.865180407
#> 36  036  0.346785362 -0.068290260
#> 37  037  1.708300311 -0.008371473
#> 38  038  0.730778528 -0.930800500
#> 39  039 -0.425016122 -0.437843329
#> 40  040  0.284957032  2.106224815
#> 41  041  0.315513555  2.337388799
#> 42  042  1.746297491  1.798981466
#> 43  043  1.543855690  1.219811133
#> 44  044 -0.751195297 -0.882987234
#> 45  045 -0.649326628 -0.563174495
#> 46  046 -0.735600060 -1.131840870
#> 47  047 -0.135772577 -0.997106871
#> 48  048 -0.840278633 -0.022572221
#> 49  049 -0.296876307 -0.132929763
#> 50  050 -1.475017585 -1.154103229
#> 51  051  1.853904634  0.878169362
#> 52  052  0.596590191  0.935102581
#> 53  053 -0.275346068  0.911502869
#> 54  054 -2.301765992 -2.610172509
#> 55  055  0.402320971 -0.966980342
#> 56  056  1.085366353  0.405620642
#> 57  057  0.387714633  0.897841298
#> 58  058 -0.067672207 -0.024816882
#> 59  059  0.757609252  0.015911053
#> 60  060 -0.955068970 -1.641733807
#> 61  061  0.628827519 -0.449581963
#> 62  062 -0.454508657 -0.725713245
#> 63  063 -0.538459205 -1.433076427
#> 64  064  0.716291584  0.451911338
#> 65  065 -0.573563494 -0.376696384
#> 66  066  0.456590468  0.637942805
#> 67  067  0.430482615  0.521447828
#> 68  068 -1.456435669 -0.329195040
#> 69  069  1.269333157 -0.368912585
#> 70  070 -0.853664460 -0.541270325
#> 71  071  2.556718951  1.185060809
#> 72  072  1.764351924  1.353292557
#> 73  073  0.478373481  0.665366992
#> 74  074  0.420818949  1.114537804
#> 75  075 -0.333312221  0.692223688
#> 76  076  0.032324256 -0.032094654
#> 77  077  1.344388320 -0.003486935
#> 78  078  0.809178749  0.699795145
#> 79  079 -0.220888617 -1.442108946
#> 80  080 -0.742514811 -0.178574806
#> 81  081 -0.155548822 -0.053180702
#> 82  082 -1.651477546 -0.120158632
#> 83  083 -1.651569477 -0.070181696
#> 84  084  0.583554805  1.080503278
#> 85  085  0.995273095 -1.532832594
#> 86  086  1.922995812 -0.093971797
#> 87  087  0.984745251 -0.630360135
#> 88  088  0.716777960  1.367480698
#> 89  089  0.650028451  1.152761335
#> 90  090 -0.406603042 -0.789492024
#> 91  091  1.057504305  1.943307642
#> 92  092  0.003716907  0.477577995
#> 93  093  0.624563272 -0.299023643
#> 94  094 -0.885076544  1.032056894
#> 95  095 -1.041100958 -1.686986280
#> 96  096 -0.051408973 -1.223276231
#> 97  097  0.878668422  1.110377877
#> 98  098  0.756361975  1.801364379
#> 99  099 -1.790307913 -0.981291334
#> 100 100  0.298161801  1.420433815

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001  0.00712882 -0.49368549
#> 2   002  0.63019827  0.77455709
#> 3   003  0.28776535  0.45096277
#> 4   004  0.56685050  0.51697333
#> 5   005 -0.15043307 -0.99858120
#> 6   006  0.67057029  0.87883196
#> 7   007  0.12462094 -0.79660125
#> 8   008  0.81531030 -0.46945835
#> 9   009 -3.25636130 -0.68203727
#> 10  010  0.04526999 -0.63037805
#> 11  011  0.28620991 -1.06365153
#> 12  012  0.28010779  1.21635573
#> 13  013 -0.16039869  0.51620685
#> 14  014 -2.34859792 -0.41722465
#> 15  015 -0.65378653 -0.51381250
#> 16  016  0.47733613  0.47744374
#> 17  017  0.17234488 -1.46413751
#> 18  018  0.41019387  0.37107549
#> 19  019 -0.38318911 -1.49629432
#> 20  020 -0.14486931 -1.09493519
#> 21  021 -2.16621215 -0.53170803
#> 22  022 -0.06357501 -0.17855202
#> 23  023  1.30932050  1.07606478
#> 24  024 -0.66291105 -1.37946575
#> 25  025  0.50338067  0.13862612
#> 26  026  1.30685485  0.77859850
#> 27  027  0.82246421  1.33203661
#> 28  028 -1.62761500 -0.92625628
#> 29  029  0.92788785  1.33525250
#> 30  030  1.63380963  1.37769919
#> 31  031  1.53260827 -0.08118273
#> 32  032 -0.25085418  1.08750006
#> 33  033 -1.38356608  0.05061670
#> 34  034 -0.43766168 -0.77168116
#> 35  035  0.86038919 -0.10322366
#> 36  036 -3.02628886 -1.13679754
#> 37  037  0.10599591  0.15141108
#> 38  038 -1.43130876  0.25563416
#> 39  039 -0.80525990 -0.68899056
#> 40  040 -0.91072689  0.13854448
#> 41  041  0.27326238  0.40028210
#> 42  042 -1.16607827 -2.67008841
#> 43  043  0.77062773 -0.48004993
#> 44  044  0.82989444  1.60780759
#> 45  045  1.20922211 -0.79139355
#> 46  046  0.19370868 -0.40743607
#> 47  047  0.68298532  0.86581390
#> 48  048 -1.04114003 -0.62729430
#> 49  049  1.39190536  1.37347364
#> 50  050 -0.31123464  0.09835590
#> 51  051  0.17105056  0.09733471
#> 52  052 -0.14049096 -0.58982185
#> 53  053 -0.72093880 -0.03736625
#> 54  054 -0.07844501  0.89489677
#> 55  055  0.35387766  0.52798050
#> 56  056  1.01383743 -0.88025844
#> 57  057  0.12232115  1.28513583
#> 58  058 -0.54242589 -0.33943737
#> 59  059  0.15142902  1.99672765
#> 60  060 -1.23789571 -0.81153652
#> 61  061 -0.44842309 -1.08498572
#> 62  062  1.28746140 -0.56032016
#> 63  063 -0.13758204 -0.31771202
#> 64  064 -1.15242512 -1.38281179
#> 65  065 -1.91256995 -0.40713654
#> 66  066 -0.80229747 -1.07515382
#> 67  067  1.17493289  0.91006892
#> 68  068 -0.96221411  0.04779479
#> 69  069  0.30616556 -0.94963733
#> 70  070 -0.99737694 -1.00308896
#> 71  071 -0.76088903 -0.81835885
#> 72  072  0.70414752  0.35358230
#> 73  073 -0.86092668 -0.57411711
#> 74  074 -1.38548071  0.44829323
#> 75  075  0.16452979 -1.37267673
#> 76  076  0.43503003 -0.29077818
#> 77  077  1.33191481  1.90955327
#> 78  078 -0.92052841 -1.95623654
#> 79  079 -0.90070526  0.65348746
#> 80  080 -1.35259653 -1.19505950
#> 81  081  1.13635659 -0.14823203
#> 82  082 -0.23394568 -0.81756953
#> 83  083 -0.98210913  0.51808185
#> 84  084  0.80701697  0.16856703
#> 85  085 -0.18376841  0.38301506
#> 86  086  0.49445193  0.76602314
#> 87  087  0.80097387  1.65228955
#> 88  088  0.30647920  0.83366727
#> 89  089  0.97762336 -1.69916195
#> 90  090 -1.08503175 -0.51060328
#> 91  091 -1.18586014 -0.60248788
#> 92  092  0.43380018 -1.71169842
#> 93  093 -1.12700914 -0.49435245
#> 94  094  0.57676939  1.20665320
#> 95  095 -0.79393039 -1.90685598
#> 96  096 -0.78590278 -2.14257557
#> 97  097  0.45695273  0.50044680
#> 98  098 -1.25379005 -1.14775373
#> 99  099 -1.14833512  0.04685966
#> 100 100 -0.42347541  1.28304826