Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5287681
#> Y_2 0.5287681 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1       Y_2
#> 1  0.1197036 0.2411997
#> 2 -0.4908380 0.2510451
#> 3 -0.4313659 0.6015974
#> 4  0.2384139 1.3203644
#> 5  0.2462845 0.1885035

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1          Y_2
#> 1   001  0.25121101 -0.899026150
#> 2   002 -0.05230069  0.087995912
#> 3   003 -1.59282708 -0.180637778
#> 4   004  1.99284242  1.754662508
#> 5   005  0.34884433  0.481235410
#> 6   006  0.77087641  0.525984320
#> 7   007 -0.09295573  0.507228898
#> 8   008 -0.94671791 -0.681546766
#> 9   009 -0.49190132  1.830046244
#> 10  010 -0.70307011 -0.541819475
#> 11  011  1.12868231  0.857036935
#> 12  012 -2.03403714 -1.508877250
#> 13  013  1.00300428  1.247137286
#> 14  014  0.41224474  0.729340480
#> 15  015  0.73802877 -0.738534285
#> 16  016 -0.27476110 -1.072120000
#> 17  017  1.06206472  0.342265355
#> 18  018 -0.70685949 -0.230455024
#> 19  019  0.17529626  0.839700983
#> 20  020  1.07296321  0.610020453
#> 21  021 -0.29851813  0.288139651
#> 22  022  0.77145964 -0.297873326
#> 23  023  0.22929365 -1.130040513
#> 24  024 -0.90951346 -0.293957108
#> 25  025 -0.09310663  0.237451183
#> 26  026  0.91603237  0.219654906
#> 27  027  1.32184460  0.077574798
#> 28  028  0.24115997 -0.071264525
#> 29  029 -0.53919639  0.316593618
#> 30  030 -1.54457079 -1.302196925
#> 31  031  0.35804276  1.708982567
#> 32  032  0.83401987 -0.625359346
#> 33  033  0.18687787 -0.549189426
#> 34  034  1.00174414 -0.069148803
#> 35  035 -1.18617224 -1.398099891
#> 36  036  0.23071032 -1.078023219
#> 37  037 -1.82596399 -1.930179027
#> 38  038 -0.90167214  0.453486382
#> 39  039 -1.56156937 -0.820266334
#> 40  040  0.38183359 -1.442500841
#> 41  041 -0.10146836 -0.388562119
#> 42  042 -0.79994594 -0.244407584
#> 43  043  0.08438748 -0.090315116
#> 44  044  0.79713557 -0.621039927
#> 45  045  0.21453880 -0.960630962
#> 46  046  0.18318042 -1.422992989
#> 47  047  0.20924340  0.591165851
#> 48  048 -0.05966013 -0.514967642
#> 49  049 -1.01774767 -0.123540475
#> 50  050 -0.20978239  0.112280240
#> 51  051  1.61579238  0.251928525
#> 52  052 -0.46503486 -0.649539977
#> 53  053  0.25018380 -2.528203087
#> 54  054 -0.02818074  0.009293307
#> 55  055  0.70346644  1.152723882
#> 56  056  0.18080817 -1.452142490
#> 57  057  0.62460038  0.214438964
#> 58  058  2.07544083  3.019162639
#> 59  059 -0.36925871 -0.194029997
#> 60  060  0.71054966  2.177212834
#> 61  061  0.20784943 -0.474451243
#> 62  062  0.63476084 -0.634911074
#> 63  063 -1.95139812 -1.761253343
#> 64  064  0.68988133  1.246420458
#> 65  065 -0.99280463 -0.034079331
#> 66  066 -0.58780442  0.774664682
#> 67  067 -2.40720405 -0.640891785
#> 68  068  1.60991386  0.532986498
#> 69  069  0.57920398  1.378583891
#> 70  070 -0.57123380 -1.300962994
#> 71  071  0.97224602 -0.652621019
#> 72  072  0.56322788  1.104324949
#> 73  073  0.08590816  0.208979681
#> 74  074 -0.25271329 -0.569681161
#> 75  075  0.88967480  0.411122658
#> 76  076  1.91668460  1.475090746
#> 77  077  1.07735426 -0.878826002
#> 78  078  0.11068624  0.401763393
#> 79  079 -1.53330668 -0.553582035
#> 80  080  0.66904638  1.359211352
#> 81  081 -0.71341508 -1.336577759
#> 82  082 -0.28685620 -0.446681281
#> 83  083 -0.95020079  0.053773492
#> 84  084  1.75175314  0.964984732
#> 85  085  0.15025470 -1.806141643
#> 86  086 -0.19393042 -0.461357493
#> 87  087 -0.79390364  0.340977513
#> 88  088  1.86854934  0.798518086
#> 89  089 -0.28139291 -0.037746908
#> 90  090 -0.18171803 -0.579604476
#> 91  091 -0.95403196 -1.735122813
#> 92  092 -0.34210855 -1.210408694
#> 93  093  0.93372604 -0.322029835
#> 94  094 -1.03450007 -1.543904076
#> 95  095  0.43099658 -1.168212485
#> 96  096 -1.18138215 -0.016368931
#> 97  097  1.05109197  0.894261697
#> 98  098 -0.20469969 -0.780662858
#> 99  099 -0.21498142  0.175080160
#> 100 100  0.11213083 -0.128971105

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1          Y.2
#> 1   001  0.84618219 -0.737583845
#> 2   002  1.11810640  1.093350385
#> 3   003 -0.26077684 -0.196633084
#> 4   004  1.21560057  0.682752642
#> 5   005  0.56590317 -0.357249651
#> 6   006  1.34088595  2.496463350
#> 7   007  0.58285781  0.764243308
#> 8   008  1.15476350  1.347675120
#> 9   009  1.81786168  0.268515496
#> 10  010  0.40879886  0.006008847
#> 11  011 -0.49638471  0.997379657
#> 12  012 -0.14558417 -0.245289179
#> 13  013 -1.27641749 -0.412958797
#> 14  014 -1.03666389 -0.237223967
#> 15  015 -0.33950413 -0.180352332
#> 16  016  0.49369817 -0.485019102
#> 17  017  0.14190965  0.743005162
#> 18  018 -0.68987216 -0.632327754
#> 19  019  0.75766060  0.495433923
#> 20  020 -0.53347834 -0.786963654
#> 21  021 -0.80748967  0.133953693
#> 22  022  0.53072398 -0.748754338
#> 23  023  2.90385129  2.054124269
#> 24  024  2.73828620  0.267866379
#> 25  025 -0.13511743  2.744837347
#> 26  026  0.30272757  0.515436350
#> 27  027  1.65637929  0.646674795
#> 28  028 -0.44098503 -0.130176567
#> 29  029  1.31465962  0.780558875
#> 30  030  0.18200813  0.414749429
#> 31  031 -0.40667647  0.047305907
#> 32  032  0.52398370 -0.053795522
#> 33  033 -0.85350033  2.064757496
#> 34  034 -0.64830845 -0.597440910
#> 35  035  0.70428781  0.872489767
#> 36  036 -0.47554364 -1.142025438
#> 37  037  2.13280188  2.130952414
#> 38  038  0.61454464 -1.926958823
#> 39  039  0.36079290  1.174589566
#> 40  040  2.03991141  2.085911509
#> 41  041 -0.65749863 -0.189602401
#> 42  042 -1.15744606 -0.767809559
#> 43  043  0.79188297 -0.167791408
#> 44  044  0.06800390  0.364616064
#> 45  045 -1.18223401 -1.058226034
#> 46  046  1.09353624  0.528836583
#> 47  047 -0.03329666  0.283855978
#> 48  048  0.92362797  0.676340213
#> 49  049 -0.28523959  0.344676005
#> 50  050  0.56396841 -0.153846430
#> 51  051  1.53833708  1.012939835
#> 52  052  0.08045723 -0.671073946
#> 53  053 -0.06217107  0.548385279
#> 54  054  0.15215293  0.301794999
#> 55  055  0.24317700 -1.281782142
#> 56  056 -1.92886812 -0.286420971
#> 57  057  0.93114201  0.538732880
#> 58  058  0.42279294  1.497597531
#> 59  059 -0.73962125 -0.321638811
#> 60  060 -0.64384836 -0.533309834
#> 61  061  1.57693277  0.826992919
#> 62  062 -0.32506613 -0.141317176
#> 63  063 -0.74038015 -0.686643589
#> 64  064 -1.07978087  0.439089589
#> 65  065  0.41406245  0.382187242
#> 66  066  0.51406936 -0.419865631
#> 67  067  0.61898876 -0.007245226
#> 68  068 -0.60387567  0.355338393
#> 69  069 -1.74752070 -0.438689480
#> 70  070  1.83889022  1.350493064
#> 71  071  1.50263389  1.207069326
#> 72  072 -1.20347392  0.409913929
#> 73  073 -0.06694321  1.106936334
#> 74  074 -0.72034780  0.337695270
#> 75  075 -0.82596114 -0.647471142
#> 76  076  0.23955766 -0.944028845
#> 77  077  0.67565150 -0.260701645
#> 78  078 -0.76669007 -1.192314149
#> 79  079 -0.89706368  0.207566056
#> 80  080  0.65750532 -0.788899418
#> 81  081 -0.54418477 -0.043272555
#> 82  082 -0.38477151 -1.043262906
#> 83  083  0.37804320 -0.965849412
#> 84  084  0.76542434  0.248094891
#> 85  085 -0.03441943 -0.088938528
#> 86  086 -0.07240946  0.495559579
#> 87  087  0.20995020 -0.117200012
#> 88  088 -1.28896903 -0.885856031
#> 89  089 -0.66836474 -0.263456431
#> 90  090 -1.21281420  0.450404099
#> 91  091 -2.73431803 -0.059633066
#> 92  092 -1.05872250 -0.679577067
#> 93  093 -1.62213472 -0.629035341
#> 94  094  0.82372016 -0.630574036
#> 95  095 -0.89688506 -1.566059235
#> 96  096 -0.56496702 -0.499232900
#> 97  097  0.14713491 -1.424019200
#> 98  098  0.31609701 -1.035036911
#> 99  099 -0.04873628 -0.443008341
#> 100 100  1.04346411  0.547991483