Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.4611069
#> Y_2 0.4611069 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1        Y_2
#> 1  0.1618664  1.0671961
#> 2 -0.6628268 -1.5893371
#> 3 -0.3493015  0.3216442
#> 4  0.4203325  0.4949395
#> 5  1.2849941  0.8532293

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID           Y_1         Y_2
#> 1   001  7.586808e-05  0.01014169
#> 2   002 -4.839103e-01 -0.02134952
#> 3   003 -7.488629e-01 -2.46303595
#> 4   004 -9.762606e-01 -1.39861839
#> 5   005  2.267935e+00  1.70502688
#> 6   006 -8.623971e-01 -0.10984917
#> 7   007  6.244992e-01 -1.04793667
#> 8   008  4.362307e-01 -0.79628970
#> 9   009  2.196178e-01  0.29294399
#> 10  010 -1.160901e+00 -1.67819199
#> 11  011  9.499345e-01  0.80575382
#> 12  012 -8.223945e-01  0.85946017
#> 13  013 -1.610033e+00  0.85793532
#> 14  014  1.265288e-01  0.56181872
#> 15  015  7.271177e-01 -0.28581060
#> 16  016  1.878546e+00 -0.20769750
#> 17  017 -1.705535e+00 -0.32058850
#> 18  018 -1.238949e+00 -1.22421628
#> 19  019  1.684762e+00 -1.21964330
#> 20  020  1.082164e+00  0.83502434
#> 21  021  3.297751e-01  0.86108418
#> 22  022 -9.455740e-01 -0.77168010
#> 23  023 -4.666306e-01 -0.23855357
#> 24  024 -2.223021e+00 -0.50127142
#> 25  025 -3.459744e-01 -1.87314862
#> 26  026  4.732609e-01  0.14104265
#> 27  027  5.883775e-01  1.57410757
#> 28  028  2.721380e-02 -1.18560663
#> 29  029 -1.319520e+00 -0.55146119
#> 30  030  6.738841e-01  0.54947989
#> 31  031  1.368950e+00  1.44092021
#> 32  032  2.816591e-01 -1.22464517
#> 33  033  1.209926e+00  1.33658585
#> 34  034  1.746993e+00  1.85069512
#> 35  035 -1.323603e+00 -1.87378714
#> 36  036 -1.135374e+00 -0.79278871
#> 37  037 -4.611539e-01 -1.52126972
#> 38  038 -1.455950e+00 -0.92212485
#> 39  039 -1.182247e+00 -1.04039807
#> 40  040 -1.010364e+00 -0.50912490
#> 41  041  9.222534e-01 -0.46757201
#> 42  042  4.770815e-01 -0.68053578
#> 43  043  2.473183e-01 -0.39343524
#> 44  044  5.434382e-02 -0.94862861
#> 45  045 -2.964296e-02 -2.21307242
#> 46  046  4.685693e-01  0.91725476
#> 47  047 -1.272755e+00 -0.24160845
#> 48  048 -1.831813e-01  0.47866971
#> 49  049  5.818463e-01  1.77271173
#> 50  050 -1.192228e+00 -1.50729162
#> 51  051  4.923178e-01  0.54209908
#> 52  052 -8.559976e-01 -0.81732848
#> 53  053 -5.055953e-01  1.02907459
#> 54  054  3.473826e-03 -0.05056411
#> 55  055  2.021040e+00  0.22922349
#> 56  056  6.746336e-01  0.19573677
#> 57  057  4.175965e-01  0.50279550
#> 58  058  4.969807e-01  0.70344177
#> 59  059 -7.299068e-01 -1.52401393
#> 60  060  1.506865e+00  0.77325951
#> 61  061 -4.855857e-01  0.46914556
#> 62  062  2.294492e-01  1.43186243
#> 63  063 -1.087335e+00 -1.36551997
#> 64  064  2.461661e-01 -0.53953188
#> 65  065 -6.347559e-01  0.23688939
#> 66  066 -1.850003e+00  0.86485486
#> 67  067 -2.035962e+00 -1.87022455
#> 68  068 -2.828442e-01  0.61924246
#> 69  069  6.555338e-01  0.65437280
#> 70  070 -1.847189e+00  0.29566013
#> 71  071  4.679686e-02 -0.50376484
#> 72  072 -9.182737e-01  1.06967719
#> 73  073 -2.835151e-01 -1.14174489
#> 74  074  6.269342e-01 -0.20098065
#> 75  075  1.612630e+00  3.12918873
#> 76  076  1.081914e+00 -0.14955058
#> 77  077  6.193277e-01  1.08355020
#> 78  078 -8.163399e-01 -0.34727563
#> 79  079 -5.503650e-01 -0.23623612
#> 80  080 -7.787371e-01 -0.08121089
#> 81  081 -1.604951e+00 -0.26706672
#> 82  082 -6.287616e-01 -0.81524505
#> 83  083  6.681230e-01  1.02069676
#> 84  084  4.998817e-01 -0.39200042
#> 85  085  1.881474e+00  0.22182887
#> 86  086 -1.173193e-01  0.40737421
#> 87  087 -6.883003e-01  0.37590278
#> 88  088  3.210206e-01 -0.91585158
#> 89  089 -8.452491e-01 -0.36395732
#> 90  090  1.376829e+00  0.20684513
#> 91  091  3.040309e-01  1.48698320
#> 92  092  6.750654e-01  0.14993872
#> 93  093  2.299206e+00  1.20551860
#> 94  094  3.525086e-01  0.74756530
#> 95  095 -2.115915e-01 -0.30394517
#> 96  096  9.334370e-01 -0.58219109
#> 97  097  5.082283e-01 -0.58010991
#> 98  098 -1.729153e+00 -0.97427547
#> 99  099 -6.629035e-02 -0.92719652
#> 100 100 -4.456631e-01  1.36819023

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID         Y.1         Y.2
#> 1   001 -0.05742378  0.16609895
#> 2   002  1.00833496  0.67207989
#> 3   003 -0.06388522 -0.17317876
#> 4   004  0.37971631  0.89338622
#> 5   005 -0.28393832 -0.56485394
#> 6   006  0.10492722  0.37180607
#> 7   007 -1.30445924 -1.51271787
#> 8   008  0.92618055 -0.35213694
#> 9   009 -0.22257576  0.86292195
#> 10  010  1.10111425 -0.86024425
#> 11  011  0.69927189  0.83190678
#> 12  012 -0.74747041 -1.48314183
#> 13  013 -0.44634850 -0.13323237
#> 14  014 -0.15315182 -0.19709928
#> 15  015 -0.33803584 -0.16899794
#> 16  016 -1.25179538 -0.68497018
#> 17  017  1.39789633  0.39850091
#> 18  018  0.01374581  1.11308595
#> 19  019  0.17522787  0.80328128
#> 20  020  0.32646870  0.68705158
#> 21  021  1.32256842  0.30101791
#> 22  022 -0.35398927 -0.91215824
#> 23  023  0.57749049 -0.92101923
#> 24  024  1.70253963  1.60738519
#> 25  025  0.66805402 -0.45550348
#> 26  026  0.93062511  0.16832158
#> 27  027  2.20315400  1.14691840
#> 28  028  0.46504824 -1.37810403
#> 29  029  0.33555327 -0.21896921
#> 30  030 -0.27073595 -0.03752922
#> 31  031 -0.03017050  0.37876618
#> 32  032 -0.29246202 -0.09471099
#> 33  033  0.86962932  1.34748019
#> 34  034 -0.28870428 -0.20658591
#> 35  035 -0.45193804 -0.33905385
#> 36  036 -0.50176101 -0.12716898
#> 37  037 -2.40170730 -2.17385548
#> 38  038  1.00494351  0.46913713
#> 39  039 -1.25266955 -0.72511293
#> 40  040  1.36003972  0.31204287
#> 41  041 -0.71007175 -0.66258480
#> 42  042 -1.19473661 -1.12666978
#> 43  043 -0.34407011 -1.59635390
#> 44  044 -2.22774554  0.39342495
#> 45  045  1.12344416  0.61471664
#> 46  046 -1.41164362 -0.64395902
#> 47  047 -0.23536677 -0.46370254
#> 48  048 -0.95609883 -0.18125290
#> 49  049 -1.59733793 -1.04919913
#> 50  050  2.86570316  1.78049505
#> 51  051 -0.84165260 -0.67357427
#> 52  052 -0.33226591 -1.00737614
#> 53  053  2.07551860  0.90724952
#> 54  054 -0.91106339  0.03066742
#> 55  055 -0.78423535 -1.08780183
#> 56  056  1.19249528  1.35317166
#> 57  057 -0.39722541  1.03337118
#> 58  058 -1.79548218 -0.03341597
#> 59  059  0.49603332 -0.10612526
#> 60  060  0.09353129  0.12880648
#> 61  061  1.00424258  0.74310527
#> 62  062 -0.84953570  1.44631713
#> 63  063 -1.39905737 -0.95068900
#> 64  064  0.90524119  1.46825962
#> 65  065 -2.27680972  0.48068634
#> 66  066 -0.04990250 -0.85854936
#> 67  067 -0.91752181 -0.79737118
#> 68  068 -1.53500245 -0.78482401
#> 69  069  1.70829720  1.20073924
#> 70  070  0.54695951 -0.02410806
#> 71  071 -1.28812014  0.69689443
#> 72  072  0.16805879  0.67492314
#> 73  073  0.74419677 -1.26625956
#> 74  074  0.65692170  0.24307142
#> 75  075 -0.14571727 -1.85751043
#> 76  076  0.53717495  0.64690168
#> 77  077  0.53548042 -0.35616206
#> 78  078 -0.67876382 -1.80464917
#> 79  079  0.24297560  1.10656150
#> 80  080 -1.37036285  0.49938533
#> 81  081  0.91638223 -0.82610949
#> 82  082  0.51247439 -0.26313504
#> 83  083  0.18876663  0.04759880
#> 84  084 -0.11086076  0.78944635
#> 85  085  0.27161581 -0.56183674
#> 86  086  0.38876518  0.93340714
#> 87  087 -1.10317513 -0.62312041
#> 88  088  1.25339369 -0.15996331
#> 89  089  0.13401461 -0.80766814
#> 90  090 -0.86416756 -0.26261014
#> 91  091  0.42187502  0.83154713
#> 92  092 -1.56509452 -1.82933278
#> 93  093 -0.79243712 -0.63125192
#> 94  094 -0.20169657  0.18941306
#> 95  095  0.26767796  0.11517298
#> 96  096  0.09236181  1.58265903
#> 97  097 -1.25504972  0.34476142
#> 98  098 -0.80740799 -0.37213046
#> 99  099 -0.65139487  0.71162164
#> 100 100  0.38945830 -0.03646406