Draw multivariate random variables

Draw multivariate random variables

Usage

draw_multivariate(formula, sep = "_")

Arguments

formula

Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.

sep

Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.

Value

tibble

Examples

library(MASS)

# draw from multivariate normal distribution
dat <-
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = 500,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))

cor(dat)
#>           Y_1       Y_2
#> Y_1 1.0000000 0.5250922
#> Y_2 0.5250922 1.0000000

# equivalently, you can provide a prefix for the variable names
# (easier if you have many variables)
draw_multivariate(Y ~ mvrnorm(
  n = 5,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>          Y_1         Y_2
#> 1 -1.3734852 -1.21141222
#> 2  0.9245315  1.15082980
#> 3 -1.3901404 -1.36994836
#> 4 -0.8468183  0.89200715
#> 5 -0.5804396 -0.01040421

# within fabricate
fabricate(
  N = 100,
  draw_multivariate(c(Y_1, Y_2) ~ mvrnorm(
    n = N,
    mu = c(0, 0),
    Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
  ))
)
#>      ID         Y_1         Y_2
#> 1   001  0.57681745  0.76612493
#> 2   002 -0.20779450  0.55738004
#> 3   003  0.16586270  0.19825575
#> 4   004 -1.30181061 -0.72932431
#> 5   005  0.54826067  0.81466544
#> 6   006 -0.24570418 -0.25629264
#> 7   007 -1.40375401 -1.08192748
#> 8   008  1.27700701 -0.93194985
#> 9   009 -0.28369835 -1.77127753
#> 10  010 -1.75744456 -0.53603780
#> 11  011 -0.80186889 -0.31360037
#> 12  012  0.67371559 -0.34220786
#> 13  013 -0.18160433  0.26282324
#> 14  014  0.97924296  0.21443266
#> 15  015  0.79005245  0.26529258
#> 16  016  0.27898106  0.17363193
#> 17  017  1.51949727  1.53136051
#> 18  018 -0.65722947 -0.42423481
#> 19  019 -0.20382311 -2.39062651
#> 20  020 -0.16235004  0.37537519
#> 21  021 -0.41832334  0.48901283
#> 22  022  0.19010485 -1.09345595
#> 23  023 -0.79702992 -1.01714502
#> 24  024  0.46647244 -0.17237032
#> 25  025  0.98413231  0.38377176
#> 26  026  0.10823121  1.45166910
#> 27  027 -1.06420334 -0.43263645
#> 28  028 -0.94865013  0.84161847
#> 29  029  0.01692090  0.18529083
#> 30  030  0.32342574 -1.83964339
#> 31  031 -1.66356734 -0.91920482
#> 32  032  0.46199679 -0.31890218
#> 33  033  0.68504830  1.25570804
#> 34  034  1.51809049 -0.15546236
#> 35  035  0.79710951  1.36684207
#> 36  036  1.24412195  0.34628135
#> 37  037  0.30603647 -1.40824396
#> 38  038  0.77021167 -0.35597934
#> 39  039 -0.74009321 -0.28731694
#> 40  040  0.87869234 -0.70000656
#> 41  041 -0.31357187 -0.87635864
#> 42  042  0.61822868  1.50845410
#> 43  043  0.90235990  0.27663443
#> 44  044  0.28777194  1.05282143
#> 45  045 -0.55802456 -1.22979741
#> 46  046  0.47711855  0.28902638
#> 47  047  1.33193632  0.82199105
#> 48  048  1.50008389 -0.20966640
#> 49  049 -1.11453571 -2.79290491
#> 50  050 -0.64027804 -1.13157690
#> 51  051 -0.16515147  0.55744239
#> 52  052 -0.06961951 -0.56886448
#> 53  053  0.92999487  1.01369535
#> 54  054  1.39561752  0.91887456
#> 55  055  1.07324831  0.78299666
#> 56  056  0.32380679 -0.10453864
#> 57  057 -0.77353029  1.00860408
#> 58  058  1.69923553  0.58708996
#> 59  059 -1.19290087 -2.76207142
#> 60  060 -1.15599554 -1.51079268
#> 61  061 -0.70244960  0.23752475
#> 62  062 -0.09717474 -0.14634173
#> 63  063 -0.28391990  2.08927900
#> 64  064 -1.00143496 -0.94286811
#> 65  065  1.60939033  1.19801704
#> 66  066 -0.55424705 -0.59052444
#> 67  067 -0.76902688 -1.54260288
#> 68  068  0.40215393 -0.81267600
#> 69  069  0.41747468  0.88251845
#> 70  070  0.63580446  0.95981566
#> 71  071  1.74075692  0.06585336
#> 72  072 -0.79245882 -1.04494801
#> 73  073 -0.05793478 -0.65197616
#> 74  074  0.55786458 -0.61156675
#> 75  075 -1.14856056 -1.78303932
#> 76  076 -0.14040864 -0.49373158
#> 77  077 -0.59088963  0.63789024
#> 78  078 -0.49548431  1.73501782
#> 79  079  0.36509183 -0.16202099
#> 80  080  1.33202114 -0.92341189
#> 81  081 -0.61767364 -0.54521697
#> 82  082  0.65053589  1.80471019
#> 83  083 -0.80424839 -1.37431149
#> 84  084 -0.71363548  0.88876784
#> 85  085 -0.57950775 -0.09484206
#> 86  086  1.24640788  0.02347112
#> 87  087 -0.77458050 -2.62326349
#> 88  088 -0.93753594 -0.49594632
#> 89  089  0.05928909  0.76640163
#> 90  090  0.93593867  0.76614015
#> 91  091  0.03650788 -0.52783355
#> 92  092  0.85284960  0.43742357
#> 93  093  0.76342240  1.20979248
#> 94  094 -0.17405616 -0.85073408
#> 95  095  1.76318083  0.52082749
#> 96  096 -0.30676445  0.15315810
#> 97  097 -0.87117143 -1.01008159
#> 98  098 -1.14196125  0.19330997
#> 99  099  0.50960837 -0.15737383
#> 100 100 -1.13538043  0.05212807

# You can also write the following, which works but gives less control over the names
fabricate(N = 100,
Y = mvrnorm(
  n = N,
  mu = c(0, 0),
  Sigma = matrix(c(1, 0.5, 0.5, 1), 2, 2)
))
#>      ID          Y.1         Y.2
#> 1   001 -2.679023665 -1.07703086
#> 2   002 -0.084577459  0.13644527
#> 3   003 -0.098416680  1.25219818
#> 4   004  1.311028745  1.20831387
#> 5   005 -0.885931009 -0.40637357
#> 6   006  0.101310961 -0.78204261
#> 7   007 -0.780895579 -0.82093911
#> 8   008 -0.353533888 -0.06625332
#> 9   009  0.988872029  1.34844536
#> 10  010 -0.741145206  0.03267785
#> 11  011  0.347522935  0.28512075
#> 12  012  0.092639199 -0.39330985
#> 13  013 -0.311328307  1.02816846
#> 14  014  0.822165473 -1.03437428
#> 15  015 -0.569083605  0.08418260
#> 16  016 -1.219273005 -0.09207947
#> 17  017  1.475565999  0.91583503
#> 18  018 -0.103612182  0.02037395
#> 19  019 -0.487304160  1.22180676
#> 20  020  1.652905970  0.86206389
#> 21  021 -0.692518925 -1.13721779
#> 22  022 -1.124750323 -0.64933923
#> 23  023  1.707689909  2.19828542
#> 24  024 -0.288327357 -0.24458473
#> 25  025  0.677410837  0.20946125
#> 26  026  0.175830287  0.50353178
#> 27  027  0.231107398  0.29354921
#> 28  028 -0.184547995 -0.44336929
#> 29  029  0.591250520  0.15093902
#> 30  030 -0.650307531  0.86217909
#> 31  031  0.478180331 -0.91070085
#> 32  032  0.066218237 -0.45410425
#> 33  033 -1.477628302 -2.22278621
#> 34  034 -0.920144028 -0.47055275
#> 35  035 -0.390463056  1.37678144
#> 36  036 -0.969594850 -0.39503742
#> 37  037 -0.006537978 -0.94031361
#> 38  038  0.370524556  1.88178017
#> 39  039 -1.314480889 -1.28717650
#> 40  040 -2.132190634 -0.61370455
#> 41  041 -0.352936694 -0.76132400
#> 42  042  0.053295333 -0.35713567
#> 43  043 -0.592078441 -1.24492488
#> 44  044 -1.196744856  0.33990358
#> 45  045 -0.069366031 -2.06222408
#> 46  046  0.152419089  1.90658883
#> 47  047  0.220008400  0.32194004
#> 48  048 -0.205243873 -0.92712660
#> 49  049 -0.495584402 -1.12381548
#> 50  050  0.341543939  0.36552304
#> 51  051  0.207045880  0.09154801
#> 52  052  0.410305331  0.75272544
#> 53  053  0.218455940  0.71741228
#> 54  054  0.846816943  0.24998534
#> 55  055 -1.479132866 -2.32706683
#> 56  056 -0.302292998 -1.04958702
#> 57  057 -0.108008727 -0.88898072
#> 58  058 -0.455276872 -0.18666812
#> 59  059  0.541353037 -1.59392288
#> 60  060  0.622362820  1.10177117
#> 61  061  1.403215384 -0.11346130
#> 62  062 -3.424048156 -2.34281961
#> 63  063 -0.463136396 -0.13548947
#> 64  064  0.013710109  0.90188703
#> 65  065  1.661239387  1.68906234
#> 66  066  0.505199342  0.02800287
#> 67  067 -2.039756796 -0.46897887
#> 68  068  1.854449956  1.93639367
#> 69  069 -0.742044358  0.05578410
#> 70  070  1.661102582  0.80152156
#> 71  071 -0.632936041 -1.83543636
#> 72  072  0.351236626  0.95579857
#> 73  073  0.486415570 -0.08150841
#> 74  074  0.462062174 -0.56046560
#> 75  075  0.647561108 -0.61282057
#> 76  076  0.341524843  1.20189060
#> 77  077  1.350045099 -0.93521619
#> 78  078 -0.764059312 -0.56398859
#> 79  079  0.147168992  2.21116737
#> 80  080  0.509212407  0.50417531
#> 81  081 -1.494444374 -0.36651424
#> 82  082  0.936808937  1.03423971
#> 83  083  1.610827601  1.58856856
#> 84  084  0.116491766  0.96706843
#> 85  085  0.531425617  1.27144886
#> 86  086 -0.135057520 -0.14567665
#> 87  087  0.120831633 -0.92705090
#> 88  088 -0.747707557  0.08267465
#> 89  089 -0.605725089 -0.59426497
#> 90  090  0.247823335  0.49077251
#> 91  091 -0.290089511 -0.56732352
#> 92  092  0.931579511 -0.80703157
#> 93  093  2.459534486  1.23124613
#> 94  094  0.047831821  1.14234752
#> 95  095 -1.660789551 -1.31455773
#> 96  096  0.746979319  0.05003573
#> 97  097  0.381916685 -1.33960223
#> 98  098 -0.505718897  0.38530424
#> 99  099  0.734317810  1.10998921
#> 100 100  0.478232962  0.26430283