Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

total_sd
A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -0.923714135  0.435003908  0.106210040 -1.735893566 -1.958475895
#>  [6]  2.275082144  1.191685202 -1.402936631 -0.758941747 -0.309205013
#> [11]  2.034923338 -0.098451338 -0.649870816  1.620310335 -1.032563890
#> [16]  0.245438505  0.355238609 -2.489406298  0.562725031 -0.211194617
#> [21]  2.132610494  1.146514642 -0.955026236 -0.418745937  0.676513100
#> [26] -0.007119184  1.800087592  0.610384029 -1.475951551 -1.022859022
#> [31] -0.326315343 -0.673093572 -1.482609679  1.387075232 -0.340385257
#> [36]  0.672914897 -1.040765876 -2.594453263  1.325106851 -1.131589571
#> [41] -0.040799758 -0.769328670  0.281551529  0.725096126  0.585886481
#> [46]  0.991110980 -0.466526406 -1.332348508 -0.342500791  1.039042307

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)

#>  [1] 13.299697 10.288054 10.834659  8.264146 10.468560  8.960213 13.866706
#>  [8] 10.866341 11.284395 10.557844 11.514721  9.743246 10.086395 11.471200
#> [15]  7.540914 14.385755  7.809013 14.704079 13.189902 10.686461 12.484905
#> [22]  4.847580  7.899183  9.321026  4.744979 10.600605  9.600604  3.659792
#> [29] 10.752976 12.335599 10.519998  8.468204  9.554093 14.985876 11.051240
#> [36]  5.531273  9.300151  9.328708 13.251965 10.169315 12.695408 14.727348
#> [43]  9.737692 12.680528  5.722861 11.638553  8.730733  9.549608  9.360396
#> [50] 16.971088

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)

#>  [1]  -8.76142121  -0.03689812  -2.60114537  -9.80672990  17.66701314
#>  [6]  13.79742258   8.11359821   8.56748350   3.40382099   1.52221825
#> [11]   4.67326587  -1.56065241   0.37950131  -4.08098747  18.29684916
#> [16]  12.51647731   1.93788573  17.19810471 -18.58384830   6.60416788
#> [21]  11.31721723  -7.78151737  -9.02524773   8.85107331  20.11367405
#> [26]  -3.40752889   3.48662191  15.37626552  -4.40897647  10.23191138
#> [31]  -3.87786335   2.40783467  16.86369154  -1.15043095   4.55570329
#> [36]  11.82314240  -6.02287110   4.10399538  -0.38437187  14.86931108
#> [41]  26.55104398  -0.66631572  -0.65478131   5.33664238  17.90950753
#> [46]  20.99272636  -2.42232922  -2.56794318  -4.93321541   6.83487610

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)

#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared

#> [1] 0.1641184