Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.05516529 -0.20478948 -0.61591382 -0.15987000  0.18500271 -1.40482826
#>  [7]  0.31392824 -1.26542330 -0.18325517  2.70224125  1.41666411  2.65708943
#> [13]  0.51168477 -1.83580420  0.98038331  0.05504120  0.64536654 -1.62515592
#> [19] -2.07315746  0.31987564 -0.14095275 -0.19039007 -0.28338602 -0.35958527
#> [25]  1.83134580 -0.59098995 -1.40784773  0.81112386 -1.68905832  2.11608312
#> [31] -0.94096529  1.18736355  1.43652153 -0.95011348  1.73352137  0.33425695
#> [37]  1.39494487  0.27122350  0.56958007  1.94854812 -1.06286595  1.29368539
#> [43] -0.19228589 -0.66694713 -0.99959126  1.78429076  2.01351575 -0.20547957
#> [49] -0.11713117  3.01085083

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.1328765  7.3298730 13.2909257 -0.1324077  6.7178285  8.6981716
#>  [7] 10.4763234 11.4534964  8.2180878 13.7121577 11.2884569  5.9711245
#> [13] 12.9882269  7.2875012 15.3919755  9.7205540  8.1346900 13.3227748
#> [19] 10.7286107 16.9337168 15.3486826 10.2871001 15.3684394  4.8934554
#> [25] 17.3028225 14.0668798  9.1276178 11.3011395  9.8412253 12.5689899
#> [31] 12.5405831  5.5756569  9.8969645  5.8785308 17.1395816  9.3358094
#> [37] 10.8871544  9.8024692  6.6551383 13.3326453 16.3852577 10.5762391
#> [43] 14.0410744  4.8738688 20.7800462 12.5003319 10.9320374 10.0047184
#> [49]  4.3022851 12.4452245

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  10.7793983  19.2944158 -20.8363603   4.4156369 -19.6008757   3.2402604
#>  [7]   4.7093542   6.8631276  14.5472447  -2.3669759 -11.4227835  -3.7194222
#> [13]   2.4893089   0.7339301   5.0417598  13.4341223   5.7001530   5.2578661
#> [19]  -2.3692158 -11.1913973  13.4962545  -3.7552434  14.6782067   7.5007962
#> [25]  -5.6380384 -14.5026659  -0.6603817   5.9548589  -4.6655133   9.0193696
#> [31]  -4.9255986  -1.2405597  15.9719895   1.1147107  -2.4544283  -5.8324783
#> [37]   1.0735871   5.1625836   1.5826336   3.0360456   1.9173563   3.7544649
#> [43]  15.6145863  -6.5506758  13.6876425   6.3491724  11.8590543  -8.1469650
#> [49]   9.8392849   5.5723578

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3923251