Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.33713712 -0.14757369 -1.46538915  1.11979530  1.42337543  3.13922999
#>  [7] -0.69839881  0.48131913  0.45145035  2.81613169  2.11017616  0.33920939
#> [13] -1.27549846 -0.40462271  2.77356373  0.74922106 -0.05374648  1.54976621
#> [19]  1.54189953  2.08895850  1.04002355  0.94048287  0.22541402  0.79632236
#> [25]  0.82884784  0.99199273  0.85391900  0.08032740  0.42650976  1.86210778
#> [31]  3.28355928  1.61847908  0.82959891  0.35699443  1.99775298  1.68690703
#> [37] -0.73177193  0.37705399 -0.19997552  3.59065302  1.07832302  0.17841568
#> [43] -0.99468051 -0.53267634  1.85548385 -0.04512996  2.39039829  0.69056024
#> [49]  0.57359289  1.21252507

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.111870 12.732955 14.530936  8.859256 17.721398 16.901766 10.622997
#>  [8] 13.952883 14.358785  9.491279  9.863931  9.503988 10.643631 14.969631
#> [15]  9.131029 14.072459  9.889323 13.875642 12.016858 11.548272 11.464507
#> [22] 11.264442 12.610707 10.737191 12.545739  7.432556  9.555915 12.522331
#> [29]  8.879293 14.330505 13.504007 15.585670 11.520959 12.018434  9.485269
#> [36]  7.622807  7.081746 18.125401  9.248260 15.000602 13.810794  6.154392
#> [43]  9.239305  6.377628 13.238222  6.964886 13.199630 18.519351 11.634724
#> [50] 16.089716

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.85072725  15.96629155   8.08284489 -14.51598604   6.36019517
#>  [6]  -4.85065162   9.05508454  -3.91245980   5.72623517   2.86128974
#> [11]  -8.45421907   1.26197223  -9.35897112  -5.79948064  14.27933744
#> [16]  16.19482339  -2.48770750  -1.34017120   4.57392605   7.11283982
#> [21]   9.32818022  -6.37571213  -2.95228922  24.85457585  -0.26713515
#> [26]   6.81156705  -0.25029081  -1.80538661   3.66539922   0.60365783
#> [31]  -1.83831231 -12.89815836   4.46300537  -0.92511560  -1.47806163
#> [36]  -6.87430950   5.54440586  13.51502909  -3.86981290   9.86836458
#> [41]   8.36127339  -9.89928094  -8.57827635   2.24233139  15.03673911
#> [46]   0.06240449   0.70284817  10.74658129 -14.09749879  -1.77437322

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1954357