Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.34524751 -0.19719489 -1.08276938 -1.41984454  1.94072686  0.23940781
#>  [7] -0.02421507 -0.38844495 -1.57369486  0.91474600  1.30275950 -0.74377716
#> [13] -0.84493659  1.24589057 -0.14263832  0.97777877 -0.22691565  0.02411572
#> [19]  0.62402582  0.81188069  0.77573550 -1.20178469 -2.09661469 -0.93666496
#> [25]  0.08687179  0.61654977 -0.85355759  0.25694405  0.05238026  2.28283372
#> [31] -1.14086953 -1.20143677 -1.12176726  1.31234243  2.19457044 -0.94506372
#> [37] -0.39651836 -1.31583624 -1.00599529  0.55111798  1.21092633  0.31774634
#> [43]  0.59841038 -1.25047884  1.70916987 -0.01313520  0.46215254 -0.07776450
#> [49] -0.22386124  1.88090772

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.031253 10.180320 10.356277  8.663710 14.253493  9.273563 10.578774
#>  [8] 12.893557  8.597725  5.961490 15.552940 10.246684 12.222774 14.658529
#> [15] 11.460995 11.362664  9.258806 15.990328 13.597765  7.919374  8.292922
#> [22] 11.118978  9.558441  8.890469  8.840900 12.317614  8.384733 11.803326
#> [29] 16.027181 11.726764 13.060095  8.495060  8.768537 12.264382  9.974565
#> [36]  9.405571 10.443008 11.198051  6.105561  5.050610 11.411449 12.474193
#> [43] 14.519032  9.294313  7.134652 12.084434  9.135789 13.717257 13.518849
#> [50]  7.054069

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -12.7940539  12.5692636  11.7395875   7.1246032   5.5042158   3.6356885
#>  [7]  -1.7925604  -7.8046885  -0.8161211  19.2930916   6.4733846   8.7570252
#> [13]  -7.2820317  -6.9237328   2.9483726  -0.6131122  -1.8213626  -0.1379841
#> [19]  -0.2304015  16.4749097   4.1377951  13.7380096  -3.9437892   4.2143794
#> [25]  22.7435074 -20.0250283 -13.8378484   7.0172291  -7.9470619   8.4672864
#> [31]  -5.4527219  -4.1007671  -2.7409709  17.8597723  13.0184319   3.5992251
#> [37]   5.8260300  -0.7221449   7.6702062   9.9894742  -4.7243395 -11.0127426
#> [43] -12.6435224   5.7528350   9.7377618  -0.9805484   5.4946341  -6.9041691
#> [49]  10.6837350   7.5838928

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4862987