Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.6752447  1.5662270  3.6739305  0.1321647  0.4365626  1.5108744
#>  [7]  0.5470953  1.9595457  1.2525446 -0.7626708 -0.4029609  0.4715296
#> [13]  2.7749431 -0.5076548 -0.2017786 -1.2605798 -0.5080720  3.3969489
#> [19] -1.4032389  2.0039958  2.0616485  0.6768485  1.8541965  1.3021957
#> [25]  0.2407957  1.2772304 -0.5710899  3.1258527  0.9820807 -0.1564006
#> [31] -0.3104918 -0.8483821  1.9233787  2.3915147  1.0997761  0.4248447
#> [37] -1.4603909  1.5590049 -0.3028037 -1.8847674  1.5460326 -0.8066747
#> [43]  4.1592154  0.1755499  0.3166070  1.7623775 -1.2237106  2.3133659
#> [49]  1.2483871  0.3236311

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.053257  9.717531 14.391091  3.433485 17.683264  6.131127  6.119779
#>  [8] 11.342914  9.592013 13.073458  6.771879  7.905499  7.236140  4.487085
#> [15] 12.460784  6.197919 10.900948  7.652847 10.922520 14.141460  6.685174
#> [22] 17.781712  7.535622  7.170464 16.161101  9.827486 16.298435  6.561964
#> [29]  8.579632 14.138363 10.676383  9.935473 10.350175  9.641099 12.101659
#> [36]  8.517601 11.453260 13.091695  7.057780  8.206985 12.700522  9.706046
#> [43] 11.894404 11.483159 15.451608  7.470620  8.933542  8.986528  6.016057
#> [50] 15.872006

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -18.2715079  -1.4066370  -6.5646940   6.3985058  -7.9912900  -7.1198356
#>  [7] -21.5659531  10.6297985  12.7718099   6.8054141 -14.6422607  -7.9230429
#> [13]  15.7826982   3.5001906  -0.8864280   9.8441009   0.2089849  -8.1126595
#> [19]   9.1657502  -9.2833749   8.6540234  -1.8202500  -0.8288152   5.2640683
#> [25]  -2.1759995 -15.4903052   1.6545313  -9.8847503   8.3272129   6.8224266
#> [31] -18.1643135  11.6665958 -22.5664159  -1.4972856  -5.0532863 -17.6114956
#> [37]  -8.5839750  22.4982983  -1.1671753  -4.1717662  -6.3064015  -7.1459746
#> [43]   5.9552105  13.5265810  -4.8111841 -13.6279055   9.6751464  -1.7817774
#> [49]  14.0800261 -13.0485235

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1020478