Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.49658788 -2.27563874  1.08438086  2.26128892 -2.57566644 -0.17432220
#>  [7]  1.79817567  0.63941123  1.42287579 -1.46951814  0.86420899 -1.07746638
#> [13]  0.33952034  2.22002617 -3.19524755  0.70626771  1.10407796  1.06731865
#> [19]  2.06832257 -2.57620973  0.92887978  0.82795757  0.08463857  0.93456386
#> [25] -1.28938136 -2.18916850 -0.38371318  0.31316616  4.33108430 -1.77166056
#> [31] -0.80876496  0.05001156  0.77537693  2.51983347 -2.62378856 -0.29752757
#> [37]  0.45139776  0.60162336  3.67624261 -3.17967496 -1.34690434 -1.83816431
#> [43]  1.92143943  0.57262719 -1.93252633 -1.89291340 -1.41950162  0.54045242
#> [49]  1.99441119 -1.76986380

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.859392  9.558953 12.332958  5.718848 15.660059 12.072473  5.245432
#>  [8] 12.011607 15.943700 12.664131  6.672966  8.946139  9.293658 17.710873
#> [15]  9.960083 16.160299  8.097214 12.615978 13.217842  9.772347 11.543370
#> [22]  3.994704 13.092671  9.360848 10.100189 11.936850  8.486500  7.373385
#> [29] 14.371785  6.935797 12.848416 11.338527  7.959017 11.051930 16.596537
#> [36] 16.712591 10.955301  8.497802  8.488950  9.560329 11.458108 11.271932
#> [43] 10.684511  9.619482 13.421236  3.262992  6.504469 13.011330 10.979082
#> [50] 12.208556

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  16.2467025   5.7267247  -0.7810647 -10.3019467  16.3536177  11.0675152
#>  [7]   9.3127853 -14.5390572 -16.8363757  18.9308825 -17.6196685   2.7048413
#> [13]  -7.8780035  -5.0481238   1.1958303   4.4219231   9.8338189  -4.2208594
#> [19]  -8.4139973  17.9685904  11.6756594  -6.4899085   3.5044036 -10.1847401
#> [25]  -6.4707486   9.5348906   1.3437890 -10.9181840  -4.4142321   2.6974202
#> [31]  15.7130274  -0.6397649 -13.9112547   0.2256492   7.4399024   8.1998744
#> [37]  -2.2464105 -12.2194583 -13.6883145   7.9702360  -1.8356269   4.4432710
#> [43]  -3.6843469 -14.4975272  -3.9690207  12.2874762  -5.7593423 -10.0940526
#> [49]   0.3247055   7.6078972

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4150186