Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.3527981 -0.7786678  0.4947387 -0.8546765  2.4498275  2.6282817
#>  [7] -2.2335455  0.6792322 -2.1168992  2.1630213  0.2976107  1.3691213
#> [13] -0.6790302 -0.7250010  1.8646460 -0.3387405 -1.0997400  1.0546053
#> [19] -0.9785924  1.7681208  1.3474877 -0.2738918 -1.3911666 -0.2293245
#> [25]  0.3661229 -1.4863254  1.6350845  0.4045853 -0.1148098  0.5394565
#> [31] -0.5555147  1.7848893 -0.4051458 -2.8179603  1.1994412  1.4628901
#> [37]  0.7772120  1.3136807 -0.3818746  1.5439904 -0.1153013  0.4182997
#> [43]  0.3976205  0.3493159  0.5046031  2.1419569  0.7201551  0.5866498
#> [49] -0.1661664  1.4426057

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.695013 10.325169  7.099954 12.200751 15.442663 13.335770  9.457465
#>  [8] 11.665668 17.178310 15.921732  6.841791  8.952658 11.012609 12.009362
#> [15]  8.210753 11.789655 16.106420  6.407879  9.671740 11.259998  7.188752
#> [22] 11.129280 13.834439 15.603831 15.092692 12.515797  9.651153  6.967658
#> [29] 11.379268 11.081484  7.486035  3.453609  6.033022 14.042470 15.636041
#> [36]  8.690999  7.816232  7.289461 12.423430 13.721698  8.520395 11.089338
#> [43]  7.855374 11.273115 15.118583  4.947076 13.126986 11.115495 14.275507
#> [50]  8.435446

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   5.76875168  -2.99730859   9.91358961   0.13381035   6.19398408
#>  [6]  -8.80807115  -1.01488690  -3.74649731  11.85269911  -9.44093789
#> [11]  -5.27917011  -3.75137646  -1.19187014   9.35836015  -7.74874475
#> [16]  -6.61524527 -15.95154637  -3.78688474   1.45705325 -16.88967238
#> [21] -13.14293645   2.19784559  -6.66359094  -2.45029163 -20.55750385
#> [26]  -6.74684496  -5.78341881   0.09767821   2.45064826  -1.55879481
#> [31]   0.48572697  -0.70725106  -8.49051121  -2.76906710 -12.16547034
#> [36] -11.46633128 -13.69579116 -12.41016127  12.16683815   0.61309976
#> [41]   1.04238451  -5.01084004  -7.62686403  18.79642064 -11.48300456
#> [46]   0.94706373 -16.52830413  -3.18988194 -11.03168397  -5.95917661

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1986645