Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.72415142 -2.15422745 -2.54264941 -0.31140190 -0.99096880 -0.43456666
#>  [7] -2.99180266 -2.11019455 -1.23045243 -4.09111269 -2.29877045 -0.82497273
#> [13] -0.67601363 -0.04236357 -1.35316446 -2.06144369 -0.44438962 -2.05278500
#> [19] -0.75802772 -0.83631605 -1.88814809 -2.03745702  0.34101884 -1.71145165
#> [25] -0.27962585 -1.13186526  0.34722191  0.41413887 -0.68151026  0.69850897
#> [31]  0.30485313 -1.40979518 -0.69479073 -0.31499869 -0.04422361 -1.45980602
#> [37] -1.39122188 -3.40013542 -0.19973838 -0.47400778  0.56186902  0.63082544
#> [43] -1.90325846 -1.17718238 -2.92695010 -0.55757457 -1.91064861 -2.76786997
#> [49] -0.83473969  0.90757989

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.271920  6.637861 15.470905 13.736596  8.884323  7.046349  9.707331
#>  [8] 14.335906 10.617333 12.085701  9.212853  9.992752  9.329538  7.493894
#> [15]  8.933384  9.443337 11.304235 12.711406 12.602831  6.125872  5.245730
#> [22] 11.265095 12.649553 15.702751  8.520392  5.935755  5.048474  8.510171
#> [29] 13.721673 12.619256  9.163587  9.949004  9.913318  7.829779 10.303254
#> [36]  5.016482 11.234747 12.387806 11.710144 11.701477  7.676398  9.397147
#> [43]  5.411925 10.050311  9.236106  8.560360  3.907026  6.331448 10.196657
#> [50] 10.641958

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  11.9294159  -7.5817657   5.1451852  -2.8789250   5.7615348  -0.6292740
#>  [7]  -3.2202054  -6.7268016 -11.9278657  10.2412195  -0.5540935  -2.0117321
#> [13] -15.5459590  14.1207096   3.3777123   5.1598458  -0.7647297  11.8653855
#> [19]   9.7074321  -4.5254203   6.1467835  15.6705872 -12.9425748   2.8255505
#> [25]  14.7676187  -5.7656074  15.0038487  -5.3117760  -0.4767254  -1.8387676
#> [31] -20.3136822  -2.3186006  10.1511703   9.1235766  -0.8244540  -0.6688584
#> [37]   3.5536465   5.1562971  -1.0873652   8.1309068  -3.3452678   1.0934362
#> [43]  -7.0065354   9.3747518  19.6837255  -7.6416653  20.1067417  -0.3094011
#> [49]   7.6837642   2.2342184

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3937153