Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.13670056 -0.82314326 -0.39520576  1.05165391  2.50384326  1.17854902
#>  [7] -1.60660258  0.59567574  0.96286173  3.15857499  1.34824462 -0.56600890
#> [13] -1.11541144  2.10766753  2.38771800  1.39647748  1.54611733 -1.61682796
#> [19]  0.08813049  3.71806788  2.45201921 -0.52527358 -1.82986735 -1.08941588
#> [25]  1.52641644  1.94916328 -0.07312875  2.09348739  0.25179599  2.56839423
#> [31]  0.42280952  1.27473239  1.11229105  1.44498221  2.71039026  2.21523839
#> [37]  0.25666138  1.34609717 -0.48283675  1.59874812 -0.08563924  1.28275844
#> [43]  1.95071692  1.25132667  4.07482373  1.36895272 -0.49108695  0.60956835
#> [49]  0.74453755  3.74209030

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.8814109 14.9214260  2.0699699  8.4971932  9.6106764  4.5768944
#>  [7] 16.0673293  5.1701056 15.8599612 11.9495135 13.2239803 12.0798488
#> [13]  6.7577941 10.5914010 11.1506754  5.9921112 17.3869244 12.0167886
#> [19] 16.4926784 14.5019700  4.5455479  3.9242072  0.6591003  9.0078811
#> [25]  9.5562321  8.7795720  8.8359591  8.2094204 10.3056742 15.5578188
#> [31]  9.5958199 10.8210300 -2.9910234  9.7504517  6.4119406 10.7966141
#> [37] 10.9243188  2.7854436  5.8833829  8.7933497  5.6244356  4.6698700
#> [43]  3.4804539  9.2717405  5.6754041  8.7897495  6.4808371  3.6963445
#> [49]  6.3899825  5.5280478

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   6.14247755  12.91846489  -1.25055113  -3.39739465  -1.70946625
#>  [6]  -6.27257613  14.72344435   6.27899770  -3.86062890   8.60877890
#> [11]   3.62264092   2.62763421  -4.00414963  -3.76760811 -17.79617988
#> [16]  -3.68105761   0.85447042   3.07159952  -2.99002571  -2.57446502
#> [21]   9.81589707  -3.15664746  -8.86042173   3.78637423  -9.03587708
#> [26] -13.29540761   1.98024482  -7.16424101  -3.21517777 -10.25233705
#> [31]   5.79981954  -6.63014851  13.23978204  -9.27698279  -6.67288526
#> [36]  -9.11783045   1.12949351  -0.37367060 -11.77415904  16.03387143
#> [41]  -5.01918554  -4.76516246   4.09997820 -13.65894083  -0.04239059
#> [46]  -5.27074204   5.56359319  -4.02350387  -9.62934893  -0.14268063

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5290101