Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.08431587  2.83994701 -1.32413595  0.78008672  1.60082137  0.42084217
#>  [7]  2.57925358 -1.33378018  1.54805400  3.10432433 -0.30437161  1.08584880
#> [13]  0.05235162  0.48534731  2.19116135 -0.13676517  1.41352018 -1.69404267
#> [19] -1.48921623  2.95959125  0.50679537  0.56795243 -2.64909287  1.21807175
#> [25]  1.68307960 -0.14755872  1.17461371 -0.44795187  1.88844330  0.08005301
#> [31] -0.63444580  3.21022264 -1.93670551  0.29249465  1.86471741  0.42428840
#> [37]  2.38764831 -1.23224881  0.15064176  0.72139367  0.36559264  1.14505064
#> [43] -1.27946449  0.19347904  3.11860078  0.17401685  1.59154782 -0.67902866
#> [49] -1.03912011  1.18361225

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.173832  1.553888 11.718292 12.535179 10.999632 10.866750  3.504344
#>  [8] 10.717933 11.298425 15.144864  8.518751  6.223300  8.875481 13.055670
#> [15]  6.207944  8.689069  3.160857  7.509922  4.803238 12.670866  7.968930
#> [22]  6.281320 11.238924 11.260035 12.324719  6.136810  7.345382  9.728925
#> [29] 11.055026  7.426811  8.159707  6.069356 12.110580  6.493700 11.194674
#> [36]  9.050234  7.324092 16.554982 10.922103 10.787922  9.206969  7.080426
#> [43] 11.628746 10.428898 11.328144  8.935626  6.722094  5.759592  8.676926
#> [50]  9.919799

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -6.5552879  -5.1122181   1.9841461   4.9564954  -8.2912302  -7.4643003
#>  [7] -18.4990455   9.5073742   3.4413992   2.4714537   4.6494075  -3.0451610
#> [13] -16.4883479   3.7121704  -0.7700129  -3.8684715   2.0019970 -11.3900206
#> [19]   2.5306794   6.9360491  -4.1869223  -2.2905488  -6.2559658  -3.7415079
#> [25]  -5.1128980  -4.1430093  13.1735311  12.3072045   1.4592247   8.0813616
#> [31]  -3.7255846  -7.7680157   7.2058948  -4.8540411   9.2472599   5.8022602
#> [37]  -7.6108467  11.2444117   0.2845766   0.6975528   6.8658131 -14.2719049
#> [43]  -0.3975029  18.8372717  -2.8358499  -6.4284779 -13.6263150  -2.7504507
#> [49]  13.5126538 -13.7733320

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6405636