Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.83732069 -1.45762098  3.54772333 -1.47158079  2.42834152 -1.61725568
#>  [7]  0.94581535  2.53910864 -0.30067131  1.24324756 -1.35992996 -1.45169728
#> [13]  2.08029689  1.45851672  0.60632322 -2.13379570  0.61215149  3.69369688
#> [19]  1.17398600  0.72679025 -0.82611149 -0.80581748  2.11160954  0.46534039
#> [25]  1.28723151 -1.49675131 -0.64512764  1.95336456  0.96850111  0.38666483
#> [31] -0.02161816  0.08533589  2.43308850 -0.83423787  1.20141464 -0.31528234
#> [37]  0.77342696  2.00791626 -0.46722843  1.35700513 -0.69253706 -0.93959315
#> [43]  1.07295175  0.91016647  1.97733538 -1.11031263  0.18244621  1.68932535
#> [49]  0.04549751 -0.39792547

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.010275 14.725351  8.758276 13.920391 10.752303 11.050590 11.180150
#>  [8] 14.794992 16.157719 16.533110  8.927132  4.968253 11.465845  9.717116
#> [15] 17.533732  7.229603 14.566560  6.242402 14.070520 13.754094  5.883024
#> [22] 12.482799 10.473352  9.128973 14.407037  8.218699 11.339284 16.450156
#> [29] 10.133126 16.707217  6.851600 11.372602 13.710964 13.662104 14.781178
#> [36]  7.562942  8.527935  3.037416 20.849878 14.036459 12.258015  7.226745
#> [43] 16.250345  8.473094 14.130795  8.703726 15.085727  8.240315 13.082184
#> [50] 14.091212

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   2.9552325  -0.2675242   1.9571130  -9.2961030  -4.0964162  13.0973063
#>  [7]   4.3077496   8.0544724 -10.1966147  -2.5039310   3.2506848  -9.4211720
#> [13]  -4.2586475  -5.4050849   2.7652580   7.1940108  15.4457712   4.7723929
#> [19]   0.7550356  10.3499908  -0.7466867   1.3241064  -3.8235993 -11.2283858
#> [25]   5.8478599  -0.1815674   7.3433687 -13.8810998  14.2478769   8.2121001
#> [31]   6.3819654  -7.6715556   6.4703164   1.8862927   1.0293650   3.5081681
#> [37]  -6.9286280  -2.2713275   7.3187425  -0.3347300   6.3722487  20.3131491
#> [43]   7.5955888  -9.5974881 -12.3037931   3.9144582   2.2046266   2.5414926
#> [49]  -4.5425743   5.7826610

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5045907