Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.22706681  1.27558415  1.11092583  1.35815936  2.98732373  0.01496309
#>  [7] -0.15027580  1.09919056  1.44126968  1.30989271  1.18184198  0.41817779
#> [13]  2.07813287 -0.28568925  0.75668814  1.18878228  1.01884922  1.36671035
#> [19] -0.43097491  1.02398333  0.67539055 -0.87051699  1.87752676  0.59264072
#> [25]  0.63065784  2.05037754  0.64792214  1.08199555  0.07015579 -0.14846535
#> [31]  1.33505699  1.78276642  2.06091241  0.53893750  0.18533240  1.46357271
#> [37]  0.89742884  0.19177499  0.36727206  0.87718703  2.74093038 -0.40233920
#> [43]  1.55995234  0.47837325 -1.06792539  2.59784184 -0.21852859  0.99711650
#> [49] -0.92926501 -0.80679787

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.368516  6.074139 14.556243  6.257842 13.175762 12.263031  9.611994
#>  [8] 11.932507  6.354202 13.603718  9.873810  4.926659 13.351694 11.429378
#> [15] 11.035880  5.044832 10.388792  9.891645  7.853222 15.377966  9.099513
#> [22]  8.906748 16.892312  5.316462  9.384350 13.231323  7.297392 18.067629
#> [29] 10.164781 10.573398  9.127861 10.506456  6.506540  9.991351 10.143240
#> [36]  7.905604  8.452611 11.887804  8.824959  8.460474  9.948888 12.730929
#> [43] 12.652129  7.342580 11.334279 11.422783 10.503904  8.631516  8.029201
#> [50] 18.976683

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   4.1899333  16.0584525  -6.5549785  -7.7121632  -6.8087689  -6.7630067
#>  [7]  -6.3707802 -22.7841158   3.9586904  -0.9340290  -1.1287048   7.3594030
#> [13] -12.3695867   2.4216351   9.1032814   0.7748659   6.9256821  -3.0356452
#> [19]  16.0438127   2.9771853  -0.8477850  -7.2613402  -3.7486371   8.1234594
#> [25]  -0.8212559   8.3189995  -4.6701393 -13.5804948   3.4494002  12.7424194
#> [31]   4.9829200   6.4033807  -8.4643566  -3.9113154   2.9915513  -9.7787144
#> [37]  -7.8014610  -9.4978032  15.2726479   6.6215225  -2.7416342   8.9257282
#> [43]  -3.3532323   1.5886127  -1.2519539   0.1880839   6.0367530   0.7820639
#> [49] -12.3116018  10.1926220

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2975408