Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.695379939  0.002245070  0.891730034  0.724721270  0.014630234
#>  [6] -4.353787279 -0.512617857  1.434292578 -2.042825455  2.239416871
#> [11] -3.125161956 -1.628559236  2.098058732 -1.824048904  2.581171744
#> [16] -0.278517457 -0.008039088  0.041668915 -2.090259895  1.203061495
#> [21] -2.633599972 -1.808163505  2.678250065 -0.795586862  0.633231641
#> [26] -2.637759173 -1.875701977  0.361922883 -0.642169445 -0.041780263
#> [31] -1.451846807 -0.570266929  1.658957615  0.044305570 -1.494817087
#> [36] -2.275204177 -1.225490586 -0.873209284 -1.805941516 -0.692323656
#> [41] -0.565909579 -1.490124674 -0.074852820  1.610997233  0.583342105
#> [46] -2.961547310 -0.568294951  1.047348015 -0.927363472 -0.318163057

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  3.9576576 19.6112978  9.6097474 10.5205417  9.8124155  4.4419575
#>  [7] 12.5876299 11.0848382  9.1921962 14.5310103  1.9609931 11.2014878
#> [13]  6.1266067 10.6170446  8.6321065  0.6588928 16.4108853 10.8838448
#> [19] 12.5695676 10.7288624  6.5882459 15.1283860  4.9331108  6.0593131
#> [25] 18.3071927  8.4941432 11.6226579  9.7870448 10.4448646 11.6392782
#> [31]  4.8515176 17.9754460 11.2682434  9.9487013 10.6813982  3.7908726
#> [37] 14.6175451  4.9401132 11.4874785 10.9149000  7.4887244 16.0154449
#> [43] 11.5726625 14.3698239 13.0642503  8.7617466 12.4527848  6.6429687
#> [49]  6.7991137 16.2761202

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -12.99684958   1.91033398  -8.12063863  -2.10663463  -5.55149258
#>  [6]   4.25717168   1.39053841   2.81996951  17.26698670  -1.85491192
#> [11]  11.02002055  -7.78432247  -5.00001574   6.19121661  -6.98440673
#> [16]  -2.21115459  -8.97586580  -1.17999494  13.35504090   4.62387879
#> [21]  -1.56505897  -4.82351222 -15.00021490  16.66915156   1.52314091
#> [26]   1.90108327   5.00294853  -3.12188550 -15.73492392  -2.63990656
#> [31]   3.71814938  -8.69336132 -12.26105774  15.73902913  -2.73921804
#> [36]  -3.26094486  -4.34444629  -6.17537958  10.38082390   0.93049796
#> [41]  -5.86547201  -2.19126820  -4.33605996   7.37940787 -22.26531356
#> [46]   1.26554090   0.07678036  -2.66643422  18.30024393   0.06848845

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.121164