Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.006556418  2.478045978  1.576645165  0.128846618  1.089736141
#>  [6]  0.266428512  2.022466719 -0.029761341 -1.553542363  2.198192745
#> [11] -3.848570973  3.787258354  1.310041287 -1.269564275  1.121503695
#> [16] -0.191277821  3.264999629 -2.479046287 -1.405639976  1.421079958
#> [21] -0.320952122  4.181314829  0.780770862 -0.112630791  0.023947799
#> [26] -0.916173302  3.963499890 -0.218120401  0.132015166  1.643232635
#> [31] -1.037149983  2.545293125 -0.427208749 -0.899624611  3.625416864
#> [36] -2.115120295  3.109056098  0.532153349 -1.324658663  1.587834661
#> [41] -0.259349986  3.392553798 -1.058548604 -2.103687692  2.723774998
#> [46] -1.669948613  2.496690330  0.283473845 -1.884261822  0.578768771

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.332316  7.480850  7.584987 16.728899  9.578639  1.587120 18.067520
#>  [8]  7.492388  9.294686  8.390437  5.355856 10.448475  9.600956 12.718634
#> [15]  9.239324  9.311071 11.843633 13.457743  6.789062  9.093730 11.442803
#> [22]  8.838169  7.927508 13.543915 14.696265 10.334069 14.629429  6.807824
#> [29] 17.700079 13.571871  6.406222 11.552566 10.168552 15.309111 13.455907
#> [36] 15.257947 13.273167  3.376719  9.657753 10.138766  8.548825 10.062913
#> [43] 11.089069 13.621415  7.443011  6.273223 17.659323 11.580503  8.788005
#> [50]  5.003365

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -2.00008974   7.74757970   3.04862248  -5.60265061  -0.12094299
#>  [6]   1.64449685   8.06021679  11.70261392  -1.98209210 -10.26190646
#> [11]  -1.67006663   0.57520069  11.23578380  -2.65606435   0.96753572
#> [16]  -8.77111464   1.59822732  13.03562436  -0.05625614  -2.34563723
#> [21]  -8.61413460  18.19259759   4.73496675   9.14542219   4.34308685
#> [26]   1.61353130   2.70194691  11.67329651  -2.39288441  -1.29996350
#> [31]  10.52978501  16.42092059   0.06278880 -10.88775316 -10.56252985
#> [36] -16.13776596  -0.35481088   1.80766484   2.09487636  -4.79462466
#> [41]  11.28493623   0.06886611  11.83648953  -2.05406804   4.88525062
#> [46]  -1.56563315   2.49630432  12.90372725 -10.46657462  15.27544730

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.460891