Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.71559882  1.29582689 -0.29492464 -1.02050441  1.35305846  0.46139735
#>  [7]  1.59369717  0.62820450 -0.19957604 -0.25778715  1.65060750  2.27695340
#> [13]  0.15700048  0.69395878 -0.60739182  1.97350441  1.50402185  1.63938621
#> [19] -1.67063297  0.61790507  1.25036118  1.52444178  1.44523549 -1.24576205
#> [25]  1.68653011  0.05592716  0.82736690  2.26399435 -0.50335270  0.64034532
#> [31]  0.28913853  2.14616252  1.73906993 -0.22617525  1.18105229 -1.50194794
#> [37]  2.59139132 -0.20426813 -0.91587699 -1.31683284  0.88327016  2.60912817
#> [43]  1.09221003 -0.22885665  0.80345550  0.39515294  2.90504396 -0.52546479
#> [49] -0.87576283  1.39461081

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  0.8832103 14.3349777 10.4648560  7.7739270  9.7536162  8.1268183
#>  [7] 12.1529019  6.7522656  4.8235718  9.2278892  3.5807248 14.9388125
#> [13]  6.0027049  9.4155101 10.8326499  3.5730427 10.2018465  5.7254905
#> [19]  7.4349220  9.5983291  5.2056268  9.5954314  9.5078805  8.2849150
#> [25]  7.6996428 12.8674437  8.9194567 11.0611179  9.1632633  9.3386223
#> [31]  7.9864529 11.2454590  6.0086787  6.9236029 11.3623578  9.3816672
#> [37]  9.7309927  2.3566940  9.5444200  5.7653202  6.4839443 11.1838506
#> [43] 12.3132589 12.2096740  5.2830397 11.4655645 10.7366383  7.2595008
#> [49]  8.1497369 11.9567376

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -3.42246115   0.84217585   5.25992792 -14.70522633  -2.51375922
#>  [6]  -0.80507357  -0.05766457   1.96737018  -4.98574571   4.55007020
#> [11]   2.97783837   1.91884368  -0.42537805  -1.02223116   1.99724508
#> [16]  -7.41463763  -4.58644657   7.48905825  -2.40025011   1.57574125
#> [21]   8.86369260   4.95504854   8.83002065 -18.28081648   3.17304101
#> [26]  12.34762484  14.74066374   4.73394359  -2.05156890  -9.31064422
#> [31] -21.00764734   2.51217749  -7.45369215 -14.05812069   7.62849267
#> [36] -12.90146974  -1.40979207  -5.15585214  -5.13038447  -7.53371530
#> [41] -10.66610503 -10.27731235  20.25033733  -8.09344247  12.55083179
#> [46]   5.08902842  -8.30010751   4.76565480  -9.57610268  13.66703262

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3717159