Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -4.35268615 -0.13809946 -1.26483193  1.68483256  0.65027995 -1.88507656
#>  [7]  1.03654477  0.64078502 -0.10567653 -0.31483022 -3.46573504  0.12327214
#> [13] -0.89052402  0.73450392 -0.45056920 -2.24788726 -0.89852292  0.06063242
#> [19]  0.81067441  0.05120790 -2.27451646 -0.79844302 -0.27516531  0.22462795
#> [25]  0.57725053 -2.79110265  0.15031753 -1.00417811  2.17586762  1.44881922
#> [31] -3.87363605 -1.00080486  0.79867222 -0.48831865 -0.30867081 -1.71821980
#> [37]  0.23989114  1.37595718  2.49001607  1.01755574 -1.17539723 -0.04458282
#> [43]  0.57532184  1.02802702 -1.78099423 -1.45180351 -0.21225517  0.79588667
#> [49]  2.30762833 -0.80417290

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 15.063347 14.992547  9.996666 12.529031  3.268503 11.248345 12.325535
#>  [8] 11.997239 13.992636  9.235125  9.015820  8.169733 14.653424 15.501442
#> [15]  9.039969  6.828373  9.739132  9.678865 20.301309 11.221877  9.041623
#> [22] 12.265823  9.651532 13.055450  9.996455 13.945929 17.055817 12.849760
#> [29]  6.048428  7.996627 12.184268 12.188077 11.707904  8.429481 11.985771
#> [36] 12.845539  9.737305 10.272665 11.222804  6.137842 11.886201  9.035340
#> [43] 14.843342 13.621274 14.320587  7.895753 14.311795  7.588011  9.266370
#> [50] 10.993293

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.1287762   2.3507547 -18.5362266  14.8414851  -5.6443257   3.0845876
#>  [7]  -6.6103196   3.0709591  -7.4457027   1.5417471  -4.5192961   0.8435314
#> [13]   1.5559492  -8.1852230  -3.4891052 -10.1938382 -22.9651161   6.3385045
#> [19]  -8.8325213  -6.0302785  -7.2273552  -0.2423480  14.1783281  -9.0297430
#> [25]   3.9480583   6.3161918   9.5821854  -9.9746316  -5.3529553 -14.6155133
#> [31]  -7.0164808  -8.2293223  -6.8582379  -4.4430172 -14.8018189   1.1507292
#> [37]   0.6013875   9.1738124  -4.7002112  -9.4276460  14.4386295  -0.9223381
#> [43] -12.8857490  -5.1513209 -13.5185083   5.1891607  -4.9296233   9.7051504
#> [49]  14.3992038  -2.3920875

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2610424