Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.84541585  1.30083375 -1.29740036 -0.28451887  0.77405224 -1.16006216
#>  [7]  3.09094087 -2.63667250  0.93706660 -2.20897018  0.25134725  3.31654250
#> [13] -0.96879645 -0.05675949 -0.26049453 -0.99866309  0.38198325 -0.48048182
#> [19]  2.88918360 -1.96392778  1.58357454  0.56121950 -1.43519306  0.99280102
#> [25] -1.34341785 -0.64205142  1.68467525 -0.36662168  1.62122053 -0.87516755
#> [31] -0.07255376  2.09971515 -1.30206294  1.31299244 -1.66024269  1.06857737
#> [37]  0.94700868  0.19759412  1.09075321  0.06304216  0.87298224  0.91573033
#> [43] -2.12059356  1.47018041  1.43617445 -0.11158478  1.68630620 -1.75324025
#> [49]  0.18273211 -1.04746396

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.2507991  8.1527085 10.4828076  5.4922256 13.1439789 10.2502994
#>  [7]  8.7974004  2.9133183  3.8092245  9.5783101  6.3761771 14.0349820
#> [13] 10.8615357 -0.5894622 10.7368515  9.5479853 11.6861271  7.8380945
#> [19]  7.6508933  6.5573709  6.6823763 12.7784825  7.8087388  6.8103216
#> [25]  7.5778138  8.6184982  7.5326626  9.6552416  6.7219266 10.7430404
#> [31]  7.9798527 11.6002789 12.0556982  7.2627284 11.8966527  9.3911847
#> [37] 11.8413192  8.8465562 10.1088153  7.7905439  9.5644903 12.5359633
#> [43] 10.5420997  8.7340122 12.6103188 13.1498356 11.1246565  6.3548775
#> [49]  6.6087649  8.6474312

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.39012472  -5.34478779  16.19161991  12.75524097   0.08752354
#>  [6]   7.03634305   2.92705394  14.98773273  17.49271230   2.62194200
#> [11]   0.58591818  -2.13897886   3.30686621  -6.86370214   0.50644919
#> [16]  -6.87669664   6.30134171  11.16247462  13.72807093  -5.93467007
#> [21]   4.47734657  15.35857872  10.15995009  10.49566970   5.03633016
#> [26]  -4.03006184   5.38583978  -0.77997163   2.92050672 -15.49456158
#> [31]  13.21810057   2.15290765  -4.24968819  -3.92655605  -2.21636816
#> [36]   0.72647776 -11.77057384  21.89725897  16.09983485  -2.61204561
#> [41]  -7.27128773  17.01406464  -3.77408447  11.17155633  -3.64759053
#> [46]  -0.24857210   6.31211361   3.53839269   2.94877195  -8.15658358

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4022971