Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.59998188 -2.30148415  1.42125477 -0.47390542  2.25468248 -0.48390311
#>  [7] -0.13464029  2.05832245  0.43369793 -1.08452740 -1.40447597  0.29988390
#> [13] -0.26536728 -1.88401868  0.21407086 -0.28097955 -0.92415607  0.56108957
#> [19] -0.91146294  1.21297001  0.53853817 -0.11853491  2.48262363 -1.93412797
#> [25] -0.79906869 -1.33344127 -2.09964031  1.81361109 -0.90358239  1.55705690
#> [31] -2.22884600  1.29333362  1.40310413 -2.46728029 -0.82513557 -0.65692610
#> [37] -0.08584883  1.61032748 -1.83692873  1.11013123  0.05505555 -0.99398519
#> [43]  0.35198763 -1.56380177  0.45134612 -1.61427239  0.28966223  0.72993209
#> [49] -1.68693467  0.32958533

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.5888865 12.3244203  9.5670993 12.0858865  6.1570408 13.2832340
#>  [7]  9.6877946 18.3873380  7.3454899  5.8480822  5.3840382 13.7751843
#> [13] 13.6939148  5.2855035 -0.6774303 11.5283387 12.0237175 10.4740887
#> [19]  6.2600252  2.1892260  9.1944855  8.0563108 11.8471197  6.8905930
#> [25] 11.7764192 10.2587438 15.9218318  9.4198208 13.5258813  2.1984784
#> [31] 15.3537497  8.4669915 10.1528404  8.7869735  8.3611927  6.2306199
#> [37] 10.9007501 15.2510712 10.6378336  6.1982517 14.9880580  4.7812462
#> [43] 14.3012716  7.4705483  5.4144505 13.9098579  6.4452701 12.1302926
#> [49]  2.2135259  6.2807503

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   7.0961528 -11.6111775   3.5116716   7.1000304  19.2800709  -1.9103942
#>  [7]  -3.0602021  20.8643881  -6.5315029  13.0515698  -5.7708602  -2.0466955
#> [13]   6.6684428  10.9589104   0.9769155  -6.3415725   5.7848250  -3.1637798
#> [19]  22.7984508  -2.0872354   0.2998524  -9.7464618   5.6765539  17.4312607
#> [25]   2.7023161   6.1182818  -0.2591605  -1.0762035   7.2145193  11.2377686
#> [31]  -6.6088842 -14.1236794  -5.9222548  -2.2786517  14.3153997   0.8302535
#> [37]  -5.7983354 -21.6104603  19.3746034  12.4118182  -2.7785433  -5.4406962
#> [43]  16.5765165  -3.2355591  11.0666708  -8.1837368   5.2983644   6.6317726
#> [49]  13.2378762   8.9709796

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.214628