Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.824142532  0.561802847  2.207072968 -0.616416150  1.534320751
#>  [6] -0.116553774 -1.939966826 -0.131039808  0.077809352 -1.078771279
#> [11] -1.557699629 -0.007731942  0.330038540  1.133847468 -0.460392875
#> [16] -0.676553369 -1.350329094 -0.046863443  2.135439586 -0.344205199
#> [21] -0.113457369 -1.363030460  0.091322700  1.784323697  1.546414882
#> [26] -1.033556596 -1.086305912 -0.246143699 -0.278788164  0.150038661
#> [31] -2.110737146 -0.539650585 -1.781000617  1.729076637  1.444886518
#> [36] -0.101856432 -1.321228953 -0.886354091  1.228549451  0.668237200
#> [41] -1.472148896  0.053147869 -1.422620326  2.911631371 -0.923511978
#> [46] -0.743728513 -0.817536659  1.713811125  0.909041275  0.335143346

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  4.931774  9.787716 11.226708 10.975317 13.619933  7.380970 14.098323
#>  [8] 14.353671 10.963604  9.341005  9.616932 13.468557  8.436083  6.240847
#> [15] 10.468383  3.080582 12.324085  8.154583  9.341143 15.284361  8.954460
#> [22]  8.286609 10.624965  7.895127  8.224432  7.133261 12.852165 10.476390
#> [29] 10.836244 10.388687  4.164952 11.821530  8.204728  7.636170  9.709760
#> [36]  9.263395 14.930064 11.181534 10.829958 15.522792 12.192098  8.447133
#> [43] 11.845871  3.742722 10.689845 15.436536  6.167547  7.387117  8.513302
#> [50]  8.966824

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   4.11881958 -12.91625221  10.03139963   6.67586908  14.97553346
#>  [6]   3.50319253 -12.63758309  25.67540986  14.99083526  -6.55331524
#> [11]   6.45583323 -10.74940157   8.35556378  -0.72954139  -3.43854146
#> [16]  -9.19783614 -12.93365920  20.30826880  16.14095037  -3.47962795
#> [21]   7.58756084  -7.12697300  17.84014248  15.86716112  -4.91412775
#> [26]   5.53744860 -23.47186632  15.44272676   6.79161055   3.99419717
#> [31]  -0.88142949 -31.55873252 -16.88646670  -4.23735320 -11.44985823
#> [36] -18.40612465   0.00849423   7.47001633   2.75000524   5.75789589
#> [41]  13.37358842  -3.19257136   9.95503063  10.30819450   9.96646740
#> [46]  -4.06653518 -12.80374540  -2.10983943   3.06590446   7.16648773

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4824482