Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.84351906  0.61526935 -0.11254445  2.89984831 -1.74396495 -3.53966963
#>  [7] -0.36784581  0.07536795  4.88502511 -1.69372757 -2.46964920 -0.34099795
#> [13]  0.57857502  3.74672076 -2.70610565 -1.04751030 -0.15797765 -1.32589959
#> [19]  3.23202690 -2.24565372 -2.93527527  0.83971829  1.63354625  4.21207152
#> [25] -0.45804733 -2.28271874 -0.56187777 -2.30678408  3.75447697 -1.34803462
#> [31] -2.66396089 -1.21059714 -0.83112607  3.72541396 -0.37898182 -0.04664806
#> [37] -1.37371535  0.79868038  2.59162387 -0.30960495 -2.87044751 -0.93961292
#> [43] -1.30559188  2.48352471 -3.89231401 -2.23884230 -2.51695234  0.46069098
#> [49]  1.85114233 -1.96303854

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.2582126  7.9792617  8.4628859 11.7493909  9.2157164 11.1187793
#>  [7]  5.0440756  9.8571478 10.0861007 10.1848732 11.5949766  6.3408683
#> [13]  7.2080673 11.3359829  8.9540194 15.8823461  7.9102093 13.8321556
#> [19]  7.3116794  1.4680059  8.0487957  7.7286947 12.3758152  7.9296876
#> [25]  7.5853365 11.1637399  8.8054614 15.7833272 15.6234905 12.9314522
#> [31]  9.1585070  7.9367432 17.9738262  7.9124708 13.7814193 11.7093356
#> [37] 11.7482078 13.9482854  6.2949255  7.3684526  7.3592194  7.0436300
#> [43] 13.8464899 15.6166930 14.4877047 10.7882471  8.9116338 14.8222664
#> [49]  7.3741995  0.6745858

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -12.0026989  -4.0613909  -1.5976074 -15.4488926  -0.8197497   0.5678704
#>  [7]  11.6883105  -4.6878905 -15.1265628 -14.1314310   8.0264470  11.3891795
#> [13]  -1.8143648  -7.8718001  -8.3656390  12.0894802   8.8383293  -9.6483027
#> [19]  -8.9194091   1.6375991 -11.8019712  -1.4400098  11.0458076 -20.2888857
#> [25]  -3.7591526   2.4741682   4.1719263   1.6255460 -20.9695198  -2.9733869
#> [31] -14.2905199 -10.8066646  10.0691960  -6.2488580 -16.3736157  -6.5395157
#> [37]  -9.4689236   7.4135605  -0.9271024 -13.0802765  13.4245350 -10.8950466
#> [43]  -5.8484495 -19.0286604   9.1327245   2.0396138  -8.0449711  -6.4403840
#> [49]  -8.0004067  11.8161536

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3470497