Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.22484341  2.57512168 -1.20999796  1.04314814 -1.49921954  3.41032756
#>  [7]  2.12659925 -0.13412845  1.73281710  1.21512780  1.07781103  0.87326814
#> [13] -1.63021548  2.16776484 -1.67024346  4.01140034  0.28664701 -2.57445668
#> [19]  2.28544123 -0.05814344  0.21591048  0.58661615 -0.06853517  1.79562846
#> [25] -0.83829021  3.12872305  0.49880093 -0.86269400  3.35194353 -0.41285699
#> [31]  1.71272111  1.31457783 -1.16694945  2.28764057 -1.52304221  1.03870685
#> [37]  2.11897964 -0.42097789  3.67341589 -0.67601924  0.55527871  1.20055408
#> [43]  0.27362186  2.49312282 -0.48711501  0.81625859  0.93750406 -0.43634930
#> [49]  3.03695887 -0.25939100

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  3.935449 10.419923 11.352945 11.280510 11.445223  8.565131  6.088372
#>  [8] 10.226310  9.433192 14.493633 14.018171 11.616749 12.913949 12.082298
#> [15] 11.115072  7.331383 10.000805  8.860086 11.102741 13.973513 13.194356
#> [22] 10.123729 12.265908 12.009036  8.885652  7.438285  8.154661  8.849964
#> [29]  6.037991  9.089867  7.647086  7.796810  9.327526 14.682652 13.861676
#> [36] 13.493479  8.929757  6.109968 13.090466 12.754159 11.064338  7.627126
#> [43]  6.877929  5.356854  6.618529  5.304346 13.812085  5.746221 10.583359
#> [50] 14.675350

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -4.2952452  -5.0003104   2.8414188   3.0972290 -15.8124668   5.9173675
#>  [7]  -7.6360212  -0.3148183  -8.4401774  -6.5934671   8.7313080  -8.4325274
#> [13]  16.6466751  -1.1963311  -8.2363834  -1.0897678  12.8761806  -1.8823696
#> [19]  -2.1175403   0.2744601  16.5201535  -0.8746016 -21.2346206  -6.3028043
#> [25]   0.7132416  15.1603453  -5.4304131 -10.8976007   6.6663670  -0.8145716
#> [31]   2.1957302   8.8223898  -0.8032214   0.5608730  -7.3618317   9.9386034
#> [37]  -2.8478026 -17.9715684 -15.0362727   4.0632644   6.1837355   2.4095290
#> [43] -14.4736726  -5.5519290 -10.4982338  10.5123897   9.1281556 -16.2715446
#> [49] -17.2880627  -2.0613914

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1138017