Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.942945936  1.011772940 -0.825045751 -0.176941981  1.738382429
#>  [6] -0.630356379 -0.492880038 -1.549598784 -0.804225127  1.156170744
#> [11]  1.792244488  0.347730923 -2.087339551  0.933663565 -0.710501085
#> [16]  2.584722592 -2.544767949 -0.712489820  0.298357139  0.140279118
#> [21]  2.080849869  0.971575803 -2.173212330  0.983947145 -0.492890903
#> [26]  1.269928255  2.946672469 -1.627414003  0.006944981  1.317734461
#> [31]  1.372735972  0.348318120 -1.753305172  0.002020098  0.428538333
#> [36]  1.587272007  0.217642774 -1.526378359  0.597103431  0.655324392
#> [41]  1.260780874  0.427695310 -3.030848312  0.349506646  0.825920710
#> [46]  1.911019479  0.287962683 -2.345299721 -0.854841305  0.473831627

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.06480011  7.04435985 14.95665259 14.24595598 16.79590488  8.86661892
#>  [7] 10.11347165  7.65003400 12.54725696 10.12707022  8.16506573  6.07734898
#> [13] 11.44539749 13.49415543 10.32128131  0.01248968  7.79236417  8.43220995
#> [19] 17.87028340 14.08118994  5.15125057 11.51236250 15.64072496 11.95170144
#> [25] 14.66162242  8.53791199  5.97024124  6.57583550 13.75156033  9.07881404
#> [31]  4.82585812  6.61223328  8.96044569 16.07731108 10.87349457  6.37579627
#> [37]  7.18689880  5.09141723 14.92305890  7.61886783 14.41931572  3.00888643
#> [43] 17.12421269 13.57425715 10.24138138  9.74474086  2.69751941 10.46840558
#> [49] 16.81147986 10.99224757

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -1.5167795  -0.1436934  -8.9154991   1.5438275  -0.8740421  -1.8415566
#>  [7]   5.4308335 -13.7778476 -12.0401660  -5.0587312   4.0872211  -5.4690902
#> [13]   3.2381658  -8.1709374 -10.6467218 -11.1571912  -9.6835101   3.6464837
#> [19]  12.5713054  -7.2934771   2.2094257  -4.8744741  10.1298968   8.1755144
#> [25]  -1.9822651  12.3063859  -9.4010663  -0.7421455 -13.8065742  -3.5125502
#> [31]   0.9690224   3.9994059   7.5126895   4.4870280  -9.7848855   8.7614308
#> [37]  -5.6318846   5.5891834  -3.4219001 -19.0307330  -1.6130941   6.2685514
#> [43]   4.2919023   3.3391986  -0.2606612  -8.1964549 -19.0387124   8.8749614
#> [49]   3.2601217 -13.5360626

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4981562