Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.70576947  0.30019132  2.01511741 -0.84184358 -1.10510234  2.25005727
#>  [7]  1.45109688  1.31351184 -1.49457407 -0.75915331  0.56122432 -0.84775461
#> [13]  1.81827632 -1.41633302 -1.27948395  0.46303328  0.18010263 -0.66812954
#> [19]  0.09009543 -0.35374174  1.05890280  2.39789107  1.40219042 -0.60354205
#> [25] -0.80193007  0.74878902 -0.66703315  2.29007953  1.48699314 -2.01832098
#> [31]  1.54259718  0.20574099  2.71431130 -2.33824927 -0.05975237  2.47157402
#> [37]  1.19744990  2.48552440  0.74468182  0.06129937  0.89269759 -0.86498983
#> [43]  2.00145057  0.89804342 -1.04126254  1.42114260  1.20426434  2.97674779
#> [49] -0.15824859 -0.23619353

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.939716121 11.726081843 14.810435390 13.036995251 15.956359857
#>  [6]  4.023376648 14.074187112  6.767604193 11.944708642 16.698677282
#> [11]  4.708641357 13.713513820 10.842484424 11.577250183 11.318349697
#> [16]  8.636097875 15.541614732  9.190915272  8.909833836 18.074949063
#> [21]  5.344179052 11.285658409  7.146480995  8.852499631 10.845626067
#> [26]  4.365937083 15.504423672 11.061628437  8.726852788 14.391981976
#> [31]  4.915688086 11.612574384  9.145495506  8.681867114 15.280521375
#> [36]  4.890169171 10.030988693  9.619992831  5.022928702 14.822962403
#> [41] -0.004081005  5.803332394  9.444828063 11.329960466 13.241888067
#> [46]  0.126845724  4.489183393  3.769539443 11.156500695 14.053885985

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  13.632251802 -16.363782787  -8.301393829   3.572670516  -4.434930740
#>  [6]   1.472000289  -6.111670710  -8.213762648  -4.194087872 -18.695271129
#> [11]  -9.714653319   7.093785574  -2.567129408   4.203809645 -14.487211762
#> [16]  10.062160147   6.016774172  -2.861917348   9.091542340 -12.406453288
#> [21]   3.522643944  -9.728396034  -4.026391464  16.414972985  -8.999110101
#> [26]  -6.254132014  -4.295248891  -1.668916526 -15.622788470   8.345645225
#> [31]  12.234103097 -13.689699596   2.226350162  -8.244176162   6.168486075
#> [36]  11.186530108 -12.014582220  -8.140756192  -4.831195834  -8.965739383
#> [41]  -2.249142136  14.812169679  -2.074970679   2.460341907  -1.563702481
#> [46]  23.556594177   0.001625634  -0.054923876  -0.344910981   3.977530662

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.302371