Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.3847321435  1.0900253367 -0.7006983408 -0.8473828215 -0.0006038767
#>  [6]  2.0368462556 -1.0046598903 -1.8939517506 -1.8876263358  0.1005760765
#> [11] -0.4562889959  0.6757755357 -2.0029325504 -0.4652796503 -1.6333952351
#> [16]  0.9183426444 -0.4843935786 -1.6440842719 -1.4817644756 -1.9463822149
#> [21]  2.2594958400  1.5093181338 -0.2199525020  0.4719900438 -1.1758563460
#> [26]  0.5163666145 -0.4891333439 -1.0638111876  1.0100880691 -0.3059856979
#> [31]  1.5032632995  3.1232895147 -0.8366109002 -0.8933499753 -0.1827196058
#> [36]  2.0302337736 -0.7472437631  0.0938725164  1.4000391550 -0.5397247357
#> [41]  1.5962323381 -0.3933211728 -0.3186370089  0.1346807429 -0.9239045086
#> [46]  0.4617901776 -0.3863722070 -1.4826809252  0.0249461552 -0.2043591453

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.629771  7.375628  9.602637 11.109512  7.418873  9.313499 13.355851
#>  [8] 10.662956 11.876988  7.992000  8.221102 16.959974 10.344171 11.202502
#> [15] 11.006470  9.816748 13.068156  7.813042 15.360660 10.203549 13.930339
#> [22] 12.202263 14.239680 10.991316 10.342527 16.151706 13.728130  6.721091
#> [29] 11.185435  6.796190 14.436659 14.983837 13.417108 14.481616 12.363360
#> [36] 13.155274 14.888057 10.083741 10.643864  8.593990  9.994692  9.088985
#> [43]  6.347546 12.228448  7.995168  7.886012 10.657502 11.332988 13.967757
#> [50]  5.396868

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   7.8366910  -2.5078618  14.1233689   5.4631405   1.1276415 -13.2256159
#>  [7] -10.6996091   4.7232772   2.2320870   7.1726444 -14.1670795  -2.7346075
#> [13]  14.8020823  -5.9400740  -2.0801563  -1.9269138   2.6181240   1.5483922
#> [19]   6.9196406  -7.3062088   5.6553972   0.4950514   4.0118667  -8.0194363
#> [25]   7.8468307 -16.8371744   3.9042036  -2.4338228   7.2336293  -4.9737108
#> [31] -16.6895925   7.2759719  15.6320316   9.9033350   4.5906615 -14.1701192
#> [37]  -0.5976340  -9.2937495  -4.8337142   3.0681558  -2.1548392 -13.0995870
#> [43]   5.0176969  -3.4308117  -0.3217982  -5.7613857  -4.9847770   5.1319883
#> [49]  14.4000592  18.5101640

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6533056