Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -3.29518727 -0.66054756  0.22005538  2.29072649  0.75213662 -3.14861335
#>  [7]  0.18445411 -0.19021251  1.76608301  2.21133253 -1.73860826  0.81656521
#> [13]  1.96627808  2.09256426  1.14892260 -3.59213283  2.48647967  0.96269516
#> [19]  0.10091801  1.74771787 -2.07561903 -1.91244862 -0.33714389  1.37556506
#> [25]  0.38540088 -1.09598143 -0.80665173  1.04597781  0.58933066  1.05044515
#> [31] -2.05827064  1.64589440  1.84924458  0.01646908 -0.65575878 -0.99721739
#> [37]  0.30932579  0.30211248  1.00074254  0.62584785 -1.99749783  0.26689342
#> [43]  0.67600667  1.58754860 -0.27564749 -2.88956891  1.28512371  2.13020493
#> [49]  0.50413377  0.41156625

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.429674 11.599332  8.873025  7.401560  1.534718 11.158427  8.368864
#>  [8]  5.769383  5.935793  2.372186  4.639880  2.799255 11.898706  1.266811
#> [15]  4.157697  8.102319 12.705242 10.209504  9.389993  7.686604  4.658290
#> [22] 12.771479  7.757059  9.644605  8.228481  3.106107  5.311997  1.780615
#> [29]  6.880050  8.470277  5.780633  9.399408  5.810761  2.201648  1.741319
#> [36]  7.288276  3.752434  8.931078  1.432197  1.099743  1.888662 12.109130
#> [43]  4.645684  7.719562  7.903791  8.541879  8.273598  1.845959  6.419660
#> [50]  6.298704

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -18.33791660 -17.51988784  12.05354563   9.30404198  -5.39874917
#>  [6]  10.12857863   5.93031739  -0.06962191  12.45249318  13.78542682
#> [11] -10.67500036 -12.29798797   3.77364943  -1.19265845  12.32857249
#> [16]   4.46244000  -5.89981813  -2.45350658   6.69248546  10.15261491
#> [21]   5.32835274   7.43757465  15.51394915  -0.83377712  21.41716899
#> [26]   7.45440441 -15.51053810  25.38707414  -4.31036646  18.92168227
#> [31] -10.59989835  -3.77249643  16.98101251  11.33154203  -7.09928646
#> [36]   5.81012530  -0.73980597   7.85393794  19.12575108   7.19403177
#> [41]   2.61010924 -13.94199837  -5.41363636  -6.77809512   1.40151804
#> [46]  13.87067044   1.06699801  10.03107815  17.80693185  21.93395751

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6079303