Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.63347222  2.88858080  2.96061716 -0.10407612 -2.43605631  1.04114270
#>  [7]  0.28127578  0.21994865 -1.33847080  0.50456088  1.80911242  2.19955602
#> [13] -0.05763602 -0.68232412 -0.09543528  1.42812815  1.74793755  2.36008357
#> [19]  0.29549681 -1.91610891  0.08875712  2.31623090  3.87546371 -2.31593271
#> [25] -2.10259489 -0.63899426  2.43044360  2.31067280  0.36456390  0.88529437
#> [31]  2.50419056  1.70613420  1.93142732  0.86053662  0.31524928 -0.39359271
#> [37]  1.34755206  0.68333962 -0.60838313  0.09572670  0.04230308  1.46546219
#> [43]  1.38527037 -0.83022595 -0.71938863  1.97068223  3.00430737  1.56260384
#> [49] -0.65326089 -0.25498507

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.095954 17.906257 10.035066  7.993437 16.437871 11.683711 16.878497
#>  [8]  8.848457  9.295163  9.909776 10.153126 11.015057  7.037517 12.615371
#> [15] 11.188140  5.028651 12.581360  5.944198 15.144132 15.980709 10.675357
#> [22] 10.926572  4.494744  7.119370  8.341450 11.101982 12.609411  5.241148
#> [29]  5.249607  8.933106  8.155033 11.858607  7.911008  8.264325 11.622094
#> [36]  9.794589 11.925639 11.615038  3.170864 12.518990  9.741989 11.115038
#> [43] 10.096452 10.111550 11.243011  9.864150 13.876338  6.405902  3.001885
#> [50] 10.204788

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.9616471  15.1662756 -10.0161233 -12.1326976  -0.2889159   3.1799262
#>  [7]   7.5079874 -11.5171307   6.9320112   0.7217900   1.8733816   6.9218365
#> [13]   9.9034446   2.1617972   3.6739643  -6.5013794   9.8183800   3.3558640
#> [19]   6.1685209 -13.3405021   3.0676966  22.0405515  -8.3628545  -0.8337393
#> [25]  -4.1036810   2.5639359  15.3886342  -1.2410550  -8.0189552  -2.1601241
#> [31]  -4.0461170   9.6517403  -6.8697821  13.3249774  -4.7170250  -3.5314002
#> [37]  23.8489512   0.5406741   1.3163865   9.1968607  12.9775265  13.4883335
#> [43]  -7.5096718   5.9732998  -0.6723799   0.3843663   0.6628639   4.9561143
#> [49]   4.0009718 -12.6412979

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.299279