Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.711783492  0.007392283 -0.826630848  0.682741912 -1.943252616
#>  [6]  0.105309790  0.355410221  0.557134500  0.220459627 -1.417739893
#> [11]  1.374318985 -0.641810093 -0.227220521  1.646779407 -1.431662392
#> [16]  0.346344091 -0.311061428  1.546734058  0.808325673 -0.406017207
#> [21]  1.655156726 -1.382404182  0.613977675  1.864930525 -2.096623106
#> [26] -0.583454056 -2.488375944 -0.889311356  1.847675583 -0.991539101
#> [31]  2.296623349 -0.262580473 -1.966709545  0.174753854 -0.779866800
#> [36] -1.133341378  2.606351910 -0.942079259 -0.860296921 -1.959801616
#> [41]  0.503113316  0.181289744 -1.726970201  0.996104553 -1.075469146
#> [46]  1.083424884 -0.058945496 -0.150906591  1.608628775 -2.006906430

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.806767  9.769747 11.645348  5.948385  1.066530 11.591473 13.526888
#>  [8]  6.570872 12.677306  7.943382 10.415651 10.510702 13.214675 14.347952
#> [15]  7.948850 17.428040 10.982481  6.306292 17.847270  6.218998 11.524504
#> [22]  5.782430 15.603768 13.935404  6.254365  9.762368  9.762760 11.689011
#> [29] 12.155893  9.932893  9.886853 11.485660 11.595869 10.659859  7.057599
#> [36] 12.356770  9.960566 10.768004  6.955988  6.189459 10.959683 11.536380
#> [43]  6.142543 10.587122 11.090790 16.596721  6.844682 12.340764  4.511768
#> [50]  6.580799

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   0.02834188  -2.19706358  14.55046877  -2.40129967   0.87645164
#>  [6]   8.06194237  -2.21584170  -0.63326840 -10.47938118  -1.76925462
#> [11]   1.89012444 -10.31107299  -4.88269957  -2.82267889   0.21075925
#> [16]  -2.14690473  -2.26987853  -2.31476283  -1.99188206 -13.72146647
#> [21]  -7.95558623 -13.38108985   2.73854573  -6.11379543  -1.98065824
#> [26]  19.52820971  -1.51625527  -5.78308904   0.25165013   4.81527077
#> [31]   2.44954010   1.07482666  -1.61622919  -2.88374844   1.83456006
#> [36]   6.70671121  -2.79070309  18.68007643 -13.26692088  -9.02702533
#> [41]   8.16644291  -8.99024818  13.50915531  -9.96120714 -14.28133924
#> [46]   7.36927021   3.12911992   9.56133089  -2.50318500  -6.05534244

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2053804