Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.10776508  0.38638260 -1.60767393  0.35841167 -1.21177561 -0.33697987
#>  [7] -0.20213641 -0.11939702 -0.01413695 -1.71069557  0.18834745 -0.62299273
#> [13] -0.34567338 -1.16168847 -0.70712089 -1.05675103 -0.51684419 -1.46708861
#> [19]  0.84996196 -1.75481971  0.19512125 -0.12482092 -0.82212890 -0.94751703
#> [25] -0.96682514  0.04311174 -0.36112241 -0.11941265 -0.81113090 -0.37633662
#> [31] -1.14848286  0.91579353 -0.85074928  0.84079558 -4.09422453 -0.01548938
#> [37] -0.66825036 -0.13896283 -0.46684578 -2.11152164 -0.09117600  1.05724729
#> [43] -0.69860352 -0.63587109 -1.60183258 -1.07675029  2.19332129 -0.01973531
#> [49] -0.81121636  0.39824456

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.7167364  8.1582412  7.8212525  6.2338383 11.8666950  8.1269574
#>  [7]  9.5312272  4.9394876  3.4995065 14.7311371 11.2001030 11.7513537
#> [13]  3.8721321  0.3237693 11.0716895 11.6837185  7.2261238  6.9923552
#> [19]  6.8104988  8.2139045  8.8238834  9.3059455  6.2983300  2.3940312
#> [25] 16.1358104  8.1857754 14.9249472 11.0738335 11.0670365  8.7119269
#> [31]  7.2512850 10.5396690  4.4544498  5.4201063 11.7401947 10.4218161
#> [37] 14.1687343  6.4891594  6.9820880 10.8617532  9.6667237  2.0544355
#> [43]  8.1366466  3.6715669 14.7338833  9.3863600 13.2879954 11.7303331
#> [49]  5.0458104 14.7106122

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -6.9459990 -17.2583310   0.6914600  -6.0931888   9.2299912   7.5622959
#>  [7]  -2.0847708 -12.8728153 -19.9671026   2.4041325 -10.0356661  -5.8785855
#> [13]   0.3099158  -0.4489197   7.3446258   8.7992082  -5.7394318   0.6490443
#> [19] -22.4747507   3.4486978   4.0273213 -13.4881382  -4.1795800 -29.3482521
#> [25]  18.7075221   0.5639071 -11.8055260  -5.8123777 -12.4895384  -8.3218933
#> [31]  12.3333702   3.2124121   1.8564707 -11.1830672  -2.6664180  -1.5179438
#> [37]  -0.7907068  -4.9616466 -17.3342535  -6.5586516   0.3444540  13.2178793
#> [43]  -4.3880159 -12.9129897   3.4646614  -8.3413348   9.5795395   1.5971409
#> [49]   0.3987994  12.3197835

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.05744754