Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.77843106 -3.49864846  0.58778445  0.95326747 -1.67554253  0.81005088
#>  [7] -1.91736665  0.62778535 -0.84651158 -4.02183471  1.75635772 -1.84224810
#> [13]  2.46366709  0.05640809 -2.19598764  0.41167954 -1.17348680  2.55464188
#> [19] -0.84863757 -1.74577767 -1.01859753  0.25857849  1.06700255 -1.69271475
#> [25] -3.80509989 -0.78242123 -1.82421535  1.91022294  0.20709050 -4.01908656
#> [31] -0.18339021 -1.90891856  1.83572497  0.87501290 -0.77257862 -0.21443752
#> [37] -2.72418721  0.29648891 -0.50228244 -1.21750745 -0.98824465  0.43803424
#> [43]  1.61541325 -0.77464965 -2.42131981  0.44032230 -3.10214205 -0.82165415
#> [49]  1.47191373 -1.83205274

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.622417  9.996826  5.304978 11.514681  9.701675  7.383315  8.462847
#>  [8]  5.837182  6.461767 14.669041 13.339834  6.458357 11.190842  9.041088
#> [15]  8.738679  3.154550  9.614754 10.720995  8.547130  8.999415  6.825135
#> [22] 11.216174 12.203273 12.852384 11.025529  4.843518 15.714483  9.609303
#> [29]  9.434602 11.061750 11.334431 14.368374  9.129679  4.839806 10.623921
#> [36]  8.478327 12.460077  9.747409  9.275509  6.074262  6.558541  9.373664
#> [43] 11.299077 11.938162  9.396274  9.444188 10.151314  6.496095  5.919020
#> [50] 13.379476

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.7079804 -21.4713769   2.0529774   8.6129371  14.1244135  25.5233043
#>  [7]  -8.2349499  11.4197799  14.2242413   2.7596496  -1.6404999 -13.0665362
#> [13]   2.0047344  19.7654045  -1.1917885  -3.5751867   6.1428953  -2.9969473
#> [19]  14.7322537   3.8455747 -13.9722777 -10.8924645  17.7171512   3.3307779
#> [25]   3.5199567   3.0746256   5.1572092  -4.3790612   9.1177577   7.2884724
#> [31]  -3.3433230  -9.2398534   1.2193408   2.3404490  -5.1823694  -1.6433658
#> [37]  -8.5894919  -9.6066861  12.1617465   2.3463297  -7.4590985  -9.7898554
#> [43]   5.3417092  -1.4465502   7.0974981   6.3894069  -0.1688985   9.2606096
#> [49]   2.9667041   0.9669889

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6827769