Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.39673611 -0.21576138  2.32921515  0.60522779  0.36147921 -0.79783950
#>  [7] -0.73276284 -0.43046359 -0.29674743 -0.54394569 -0.40196903 -0.03455034
#> [13]  1.81009602 -0.35508513 -0.28700011  0.39059915  0.23518184  0.41179367
#> [19]  0.27916791 -0.44493259 -0.91352993  0.01480329  0.93208297 -0.83705069
#> [25]  1.38972161  2.06003259 -1.04339103  0.36151423 -0.72867490 -0.70996442
#> [31] -1.09904166  1.47946527  2.34268495 -1.72971155 -0.53471419  1.08955204
#> [37]  2.35110316 -0.56995971  0.27577322  0.93458681 -0.59407987  1.11855480
#> [43]  0.17584096  0.30039504 -0.90589825  0.40554742 -0.13308944  0.44256042
#> [49] -0.05535474 -2.45004887

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.196985 11.836800  8.546337  7.820752 15.950440 11.866593 14.937969
#>  [8] 12.207674  7.110883  7.252272  3.069109  8.231247 14.870160  6.073909
#> [15] 11.202454 10.142358 12.806847  8.311419  8.192696 15.212165 13.732765
#> [22]  7.488788 14.641595  6.688601 13.251405 11.788862  9.654139  8.914907
#> [29]  5.972038 17.627301 11.260051 15.288012  9.326589 12.372723 11.668367
#> [36] 12.394242  8.244683  8.189897  6.459048 16.002635 11.568037  8.297264
#> [43] 11.922539  6.670389 10.688149 13.775583  7.501488  5.831005  8.815393
#> [50] 13.635200

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.4902329  -8.1805048   6.2108389   3.4899862 -21.3645070  -0.2527036
#>  [7]  13.1065475  -9.2985630   1.2093288  10.1991291  11.1285784   7.7210327
#> [13]  -1.7171281  -5.1621607  -0.4785931  10.1136269 -17.1214910  -6.8558146
#> [19]   4.3893812  -2.8009616  10.9225369 -12.7182100   8.0741138  -6.3044392
#> [25]   1.0383458  -1.4641112  -0.2367367   7.8798198   2.8523532  -6.6014921
#> [31]   8.3095246 -10.1116280   1.9238297   4.5157863 -13.6523238  -6.7123344
#> [37]   1.4076053  -7.2222063   6.8070978 -28.8861149  -0.3478852   3.3969200
#> [43] -13.7457887  17.9018531  -5.6627484   1.9358465  -1.1823562   1.9268597
#> [49]  13.0650384 -14.7540704

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1456212