Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.7070700 -1.3105893 -0.1934792 -1.5806745  2.1694424  1.8761446
#>  [7] -1.1490661 -1.0847977 -0.8264880  0.1608217  1.3463999 -2.8307476
#> [13] -1.8101231 -1.4010276 -0.2680961  1.2378765 -0.8824362 -0.8480443
#> [19]  0.0468820 -0.2239673  1.1165331 -1.2688193 -1.5340700 -0.0801060
#> [25]  0.3993847  0.8274647  0.2676751 -1.3728134 -0.5051988  0.1556788
#> [31]  2.4525621 -1.4125817  1.3859401 -2.3368898 -1.0976145  3.6648788
#> [37]  0.5283503  0.9225317 -0.2021236  0.5106520  1.6452091 -2.4775057
#> [43] -1.1946679 -0.3663158  2.4718884  2.6892365 -0.5252571 -0.3451467
#> [49]  0.1015215  0.4919752

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.791732 11.425686 10.329133  9.943396  3.029459  8.166676 11.331651
#>  [8]  9.028346 11.267353 11.591450 13.943097 14.978646  7.211975  9.565022
#> [15]  7.613197  9.468012  4.101171  7.569581  5.098552 10.024442 11.379137
#> [22] 12.751189 13.340414 17.473851 12.031754 12.657963  8.908802 15.063531
#> [29]  9.847604 10.696670  7.137651  9.187721 10.323050 19.743521  8.285300
#> [36] 11.900484  3.328678  9.905514 10.783700 13.310939 13.835699  5.480163
#> [43] 10.785988 11.709526  7.765486  7.862029  5.061814  9.861353 13.622561
#> [50]  7.463439

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   5.1540276   1.3652805   0.4707160   2.6698415   4.2852277   1.6550270
#>  [7]   8.8575918  -2.9422770   1.9039562  -5.7050027  21.8666270   3.9390569
#> [13]   2.8768787   0.1952632 -19.9335266   5.2778667   1.0078283  -1.2872410
#> [19]  -6.7107111   8.6085952  -1.1360335   9.6505672   5.8589756  -5.4924438
#> [25]  -5.8746461  -0.9861076  13.2674155  -9.5435142  -4.8800656  -4.4238223
#> [31]  -6.1805260   7.0464582  16.9817535   0.6215060  -2.3958427  -0.6515204
#> [37]   9.3924445 -12.8685173  -5.9300320 -11.5473982   2.8075392   1.5238889
#> [43]  -7.2041197   0.5450454  -1.5274167  10.2443418   1.0319043  -0.7645125
#> [49]   3.1531372  -6.1485952

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4163562