Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.881993060  1.526796118 -0.354204581 -1.473119817  0.250948961
#>  [6]  0.172132495 -0.640468299  1.016873674 -1.138659474  0.879955918
#> [11]  1.921292663 -1.317985057  0.002492746  0.420622953 -0.474310866
#> [16]  2.144977366  0.593672923 -1.165186225 -0.058312806  0.592936964
#> [21]  0.022262443 -1.479739156 -0.224814694 -1.199039047  1.393679486
#> [26]  1.659354878  0.975856552 -3.061160489 -2.059700686  0.347090343
#> [31]  0.356126318 -0.704454855 -0.477716732 -1.291948680 -1.038438944
#> [36]  0.682986674 -0.235395397 -0.878303483 -0.769776007  0.642344062
#> [41]  1.614585131  0.180037516 -0.294995055 -0.796189458  0.500723302
#> [46]  0.981968801 -1.153473574 -1.226108440 -0.856596840 -0.094916601

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.113507  4.970985  8.664240  8.782671  7.546381 15.620207  7.535070
#>  [8]  6.255919 12.473135  4.853465  3.329967  9.646526  9.013734  7.300275
#> [15]  6.051494  9.463427  7.186783  2.722523  7.454942  5.906436  9.193124
#> [22] 14.644550  3.502939  3.635340  7.048746 10.028305 11.178918  9.934290
#> [29]  5.766320  5.646903 13.086063  1.863964  8.850344  9.358248  5.700120
#> [36]  9.856685  7.402887  9.534767  8.204739  4.925564 12.528987  9.337391
#> [43]  8.476264  7.081330  4.472636 11.701119  8.698221 11.362251  9.541793
#> [50]  3.317038

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.4471724   3.7423546   1.5898433   7.5783552   8.5392497  -3.1370289
#>  [7]  -9.5045749  -8.4424782  18.5947448   1.9428974   7.4244623  -1.6250953
#> [13]  20.0680224  13.4587240  -7.9272225  10.8006449 -12.1899750  11.4403854
#> [19] -19.3765813 -10.6517503   7.3099885 -10.5215726  20.1086963  20.7072378
#> [25]  -2.0541258  -0.4264496   3.2039359   8.3971236   9.1014578  -1.1841283
#> [31]  -2.5067340  -8.0666775   8.3968616  10.6790556   9.8695267  -0.5890748
#> [37] -14.6979459   7.8640465   9.0661655  -1.3650049   0.2163174  14.2718034
#> [43]  11.1358480  11.4642017  -1.4943661  -0.8602966 -15.8618922   3.9197145
#> [49]  18.0876713   6.5544798

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.205984