Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.632575453 -0.225501079  1.476007604  1.454616801 -0.358352553
#>  [6]  1.105393531  0.623392104  0.097992554  1.147533849  0.962010341
#> [11]  0.001255484  0.748429731  1.867589182  1.406584104  0.556572951
#> [16] -0.825000319 -0.810067736  0.715013758  1.135834405  0.062835300
#> [21]  1.265191161  1.470372015  1.578875391 -1.279753120  0.245564363
#> [26]  2.055235418  1.397248183  1.534740196 -0.020257572 -0.915818405
#> [31]  1.418816190  1.042558321  2.822673885  0.862701309 -1.058117993
#> [36]  2.611097609  1.966072602  2.261161651  0.095313717  0.298551017
#> [41]  0.121655227  0.871682241 -0.008948023  0.094337140 -1.513339600
#> [46]  0.835632933  1.559010116  1.375272871 -0.696660947 -0.125813126

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.858293 12.496042 10.669069  7.626542  5.169364 10.210924  5.069299
#>  [8]  6.006775 11.294756  9.613691  3.699476 12.964519  8.413431  8.230800
#> [15] 13.721514  3.065220 17.717118  5.491272 12.102852 10.252751 12.584766
#> [22] 14.628380  6.673179  8.128523 10.708516  4.321355  8.338946  4.337195
#> [29]  8.985532 10.845048 10.041938 12.511181  5.174744 10.699757 10.179051
#> [36] 13.646574 12.270974  9.713773  3.698955 12.845292  6.738087 14.600576
#> [43] 10.429097  4.619374  3.993471  7.889242 10.258511 16.297449  4.827410
#> [50] 10.689087

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -10.0055603  16.0975508  -3.9319068  -6.5527171  -1.5010362  -3.3437795
#>  [7]   2.1071883   2.9622663   0.7711122   2.3195841   0.5905172   2.1656185
#> [13] -16.5231361   0.4082204  -9.4629213  -6.5944426  15.7325718   5.8293982
#> [19] -16.7544634   2.7689398 -11.7391347  -7.6625745 -11.3246224   1.6811576
#> [25] -10.5481021  -7.5725262 -11.6891385  10.3859994   3.8439122  -9.2124014
#> [31] -12.1649977 -10.0142143   7.4327474  11.4119994 -12.3822578  -7.9088224
#> [37]   5.5914806  15.2294709  -2.3059881   7.3102787  -0.6099251  -5.0125546
#> [43]   7.7958179  -3.0970285 -10.0708403  -4.5751454   9.5419792  -2.9491910
#> [49]  -2.2389978  -3.8271085

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1286816