Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.60993110 -0.69787883 -1.62665389 -1.16704952  0.17163147  1.40412377
#>  [7]  0.10007591  0.37096046  0.58126371 -0.53811785  0.44048556 -0.72653365
#> [13] -0.44572586 -1.15827447 -0.77186806 -1.57473339  0.46174281  0.33659182
#> [19]  1.05343401 -0.09013012  2.30092232 -0.43468436 -0.22801549 -1.65500990
#> [25] -0.70724178  0.19914174 -0.41946957 -1.25561989 -2.77124559  0.25729550
#> [31]  0.67370468 -0.14864056  0.19643057 -1.91467926  0.56740613  1.64494788
#> [37]  0.75667023 -0.06451357 -3.05102776 -0.46420141  1.22120243 -2.70233612
#> [43]  0.74416676 -1.00246908  0.34830294  3.05359485 -0.48722369 -0.72705725
#> [49] -2.50962204  0.23704494

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  1.424724 11.289173 11.357999 11.775718 13.172550  2.304416  9.415657
#>  [8]  9.622173 12.912739 13.275835  9.536744  6.980357 11.276804  9.627218
#> [15] 16.442422  8.923400 12.434600 12.530214 12.082643 15.564990  9.514649
#> [22]  5.458498  8.066261 11.650819  9.616699 11.037288 12.888774  8.014144
#> [29]  6.916965 21.335377  9.242733  7.159481 11.026242 16.345753 13.581130
#> [36]  7.819828  9.982986 10.550960  7.134268  8.321487  4.957215 16.108143
#> [43]  7.624783 16.196533 13.739688  7.174134  6.725740 11.553662  7.167567
#> [50] 10.499179

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -10.0798595  -5.2710329  -4.5552001  14.3944510  -5.5749671   2.9106547
#>  [7]   3.0942743   7.7354681  11.7285031 -12.5725432 -11.0237320   3.2392343
#> [13]  -9.6812000  16.0692611   6.8053388   8.3712762   9.5447599 -15.9786812
#> [19]   0.4433997 -10.0608517 -11.6304695   8.9826918  -7.0599834  10.9959118
#> [25]   7.8852654  -0.9886646   7.8142840   8.1642061   9.2723306  -1.8428851
#> [31]   4.9741633  -1.6612333  -1.8894284  18.5521132  -8.6120715  -2.1866053
#> [37]   7.3464259  10.5214288  11.4890391  -0.0262309   7.0105214   4.9045604
#> [43]   1.0296845   0.4771170   5.5233586 -12.9223291  13.5095420 -20.0789125
#> [49]   7.3563689  -9.8435250

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.7443354