Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.10746819  1.89328995 -1.23937188  0.41380600 -0.40774147 -0.01506899
#>  [7]  1.20647791 -1.13102921 -0.47840554  0.33048777 -0.60917823  1.70075910
#> [13] -1.52897888  0.76391659 -0.17965685 -0.01818885  2.25723582 -2.14658628
#> [19] -0.78647665 -0.43962989 -0.28597183  0.26570143 -0.43454486  0.21461382
#> [25] -0.01834904  1.02397052  1.98212344 -0.42489644 -0.59340170 -2.21375834
#> [31] -1.54201641  1.37230371 -1.08299723 -0.07665936 -1.98913772  0.06923533
#> [37]  1.20699602 -1.42299894  1.80662392  1.31421437  1.26730581  0.35881472
#> [43] -1.77852981 -0.18449381 -1.27636136 -0.66116785  3.18051554 -0.99957409
#> [49] -0.16826928  0.17482585

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.061170 14.132313 11.906772  9.898813  5.967478  8.286276 11.737435
#>  [8]  8.832573  9.464363  7.990297 10.230325 11.468096 10.462830  8.494592
#> [15]  7.118189  7.219818  7.862018 11.022309  9.479614  6.040456 10.661182
#> [22] 12.026448 11.617520 12.486370 10.131555  8.181505  9.876734 16.775456
#> [29]  7.899891 13.766406  4.983537 11.227175 11.425505 14.244000  5.220637
#> [36]  4.055333 13.273801 16.003311  6.738180  9.300232  8.502743 10.016728
#> [43]  9.420249  6.723957  9.953474  6.092294  8.950830  6.034564 10.642033
#> [50]  1.334923

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -14.1348349  15.6224298  16.2699359   6.2051309 -13.9841866  -9.9890549
#>  [7]  -3.4059001  13.6037637   9.3328656  -3.4731421  -7.0896951  15.9050402
#> [13]  15.7509261   7.8136781  -3.8556875  -7.9590416  -2.9626436   8.5618938
#> [19]  -1.8748105   0.1175111  -3.4262403  -1.0057865  21.7416418  11.3810158
#> [25]  -5.0382059 -16.9177025  -6.1084531  10.8127030  -4.7278635  -6.0384225
#> [31]  -6.1988006  26.8545426  23.4131498  -4.2426708  -0.4013707   2.2699245
#> [37]   8.6078575  10.1381529  -6.7528137  12.2404343  -1.2513153  13.9987572
#> [43]   9.4878793   6.0006562   1.3214698 -23.7750701   3.5831157  11.6103385
#> [49]   5.8362478  -1.7205926

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4647402