Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.02779621 -0.05978775 -1.18488627 -1.05649747  0.58716474 -1.08030659
#>  [7]  1.61301604 -0.66586033 -0.77414386 -0.67281603  1.97356648  0.91118595
#> [13] -2.17955810 -2.80339320  1.08899528 -0.22802081  0.39811773 -1.16549692
#> [19]  1.56023366 -1.20695627  0.40606281 -0.49953947 -0.93699095  0.41093909
#> [25]  1.34135295  0.85580351  0.56524525 -0.05833094 -2.26375917  0.00119279
#> [31]  0.21489083  0.42384627  0.14551692 -0.16298724 -1.70803443  0.80129766
#> [37]  0.95394068 -0.97619277 -1.72210069  1.48864201 -1.28876135 -0.57299351
#> [43] -1.59992898 -1.00772887  1.02438864  2.61540674  0.28600106  0.06894780
#> [49] -0.54187792  0.68949280

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.2805289 13.7549594  0.7787121  7.9680604  2.7618477  1.8373405
#>  [7]  8.1875696  1.8080902 15.2551676  7.0378711  8.6332977  9.3122987
#> [13]  6.7643953 12.9876022 10.7298432  7.6542576  7.2592363  5.9649369
#> [19] 14.1057597 12.6863634  1.7998070 11.4529780  1.1436331  5.2336797
#> [25]  5.2014011  4.3066266  4.2387554  1.7494927 13.9518036  8.8161455
#> [31]  5.3163035  6.2991916  3.6480882 13.3633501  8.2644086 13.1741003
#> [37]  2.3444803  5.9547142 14.6569319  9.4030036  8.1340171 11.9487640
#> [43] -3.3500282  9.3794048  6.9760049 13.1918556  8.8038782  5.4976812
#> [49] 12.5733859 12.8951083

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   9.1599714 -12.4874602   1.6623523  -8.4580883   6.5899679  -1.3924700
#>  [7]  14.0460618 -14.3592133   1.0384068   5.7697416   8.6402767   2.6727843
#> [13]  18.2393987   4.5862626  -0.9976219  10.7417427  -7.7238677   1.3450034
#> [19]   6.3098874  -4.3101983  10.1660009   0.4804172   8.6376762   3.7288132
#> [25]  -1.3321017  -5.3468331   8.3759628  10.9933336  -6.5883492  -1.7055781
#> [31]   1.7378904   1.9555592  15.5032848  -6.3032580  -6.6812599   7.1533382
#> [37]  10.6068518   7.8849722   2.5809524   0.9147699  14.3564407   7.2243308
#> [43]  -6.6310737   9.1875158  -0.1857071   2.1836623  -0.3026480  -4.5270482
#> [49]  -9.5187443   0.8960577

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4785173