Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.34748895  0.34957049 -2.52466362 -0.83523747  0.38149624 -0.74766599
#>  [7]  1.05772666 -2.19697735  0.06329961 -2.27102523  2.21635000  0.72094598
#> [13] -1.11243608 -0.96966542 -1.01728576  0.58539880  2.15323412  0.47406933
#> [19] -1.82592988 -1.74799293  1.14610114  0.50783098 -1.70236859 -0.06747175
#> [25] -0.15003244  0.13934333 -0.60083060  0.63581039 -0.80692691  0.68897361
#> [31]  1.56927015  3.02449378 -1.19796713 -1.61029131  0.15707810  0.28180440
#> [37]  1.36001040 -0.20641125 -1.46137687  0.30862423  0.62200799  1.45946121
#> [43] -3.58659340 -1.43365048 -0.55498548 -0.33490656 -0.23345492 -1.83168578
#> [49] -0.63187732  0.11996936

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.879521  6.584475  8.633155 17.532190 13.512710 13.281407  9.015407
#>  [8] 10.631496 12.306522  6.288753  9.611687  5.079277  5.427490 11.402288
#> [15]  4.888098  8.715390  6.399841 13.231076 12.121058 10.581289  7.660074
#> [22]  8.497307 11.393211 11.070709  7.464069 12.965214  6.823770  4.298516
#> [29]  8.455179 11.928069 14.764025  7.418990  8.240761 12.715707  5.317327
#> [36] 11.685290 10.019505  9.431212 13.688090  6.947860  9.652829 10.922978
#> [43]  9.927608 11.361999 14.443808  5.754196  6.023179 10.612349 12.391965
#> [50] 11.001094

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -1.5666433   3.1616363  -0.6070240   2.2263915 -16.2561938   4.9540223
#>  [7]  -8.2289509   8.6290711  -0.0458487   0.1158960   2.2605502  12.1889188
#> [13]   6.6039079 -17.6841515 -10.6615982   2.2303023   8.3996412   8.5931856
#> [19]  -3.7603466  -8.6222831  11.7117572  -2.4183835  -9.3649932   4.5718128
#> [25]  -7.5182919  -1.5839441  20.3004519  20.8925631  -6.2041377  -4.6200317
#> [31]  14.9972231  12.7846932   5.4962520 -12.2496009 -14.9518538  -0.8800072
#> [37]   2.4159590   3.6679426   6.3805666   8.9264289  -0.1226163  10.7513501
#> [43]  18.2007342  -8.4997413 -13.2144333   1.1937609  15.2676913   1.3110562
#> [49]   6.0693761  -2.5798941

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3232146