Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.51606277 -1.51270137  2.66613414 -1.16094692 -1.92920757  0.81902749
#>  [7]  0.49128146  1.66948527 -1.86481506 -0.89537241  1.42064472  0.23083976
#> [13]  0.98060723 -1.68998311 -2.41308280 -1.22176481  0.60156292  0.12329481
#> [19]  0.34922206 -2.19617738  1.06727816 -0.61743560  1.45052468 -1.83694513
#> [25] -1.61514702  0.05667275  0.25474079  2.49654461 -0.61662391 -1.54299058
#> [31]  1.06803408 -0.49497754 -0.24205564 -1.78702299 -0.99501860  0.18469064
#> [37]  0.64892342  1.54476907 -2.01381078 -1.01174625  1.83707211  0.75015728
#> [43]  2.52598668 -2.02732443 -1.45109769 -0.44087293  0.37136516  1.66722684
#> [49] -2.19466193  0.08372505

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.418901  9.406929  9.259614 10.422074 18.938839 14.349989 11.220067
#>  [8]  7.116533  7.573531  8.415043 13.992841 15.885893 13.308136 13.262219
#> [15]  8.016850  4.385498 12.675935  7.870728 11.917572 13.909550  7.918714
#> [22] 15.842574 12.872469 11.620149 10.711923 11.698749 14.779219 15.558118
#> [29] 11.514309 12.087207  9.270338  6.572540 11.869089 13.674433 18.163484
#> [36]  8.744106 13.507753  9.447459 12.037753 16.202644 15.173482 12.803134
#> [43] 11.372657  9.317232 13.053005 14.359717 14.020130 14.642450 12.271698
#> [50]  9.392510

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -9.6173404 -13.9562540  20.7131033 -14.2963492   6.2404441  -6.0197064
#>  [7]  12.1647633   5.1412250  -9.3384150  -4.9668142 -12.4336660   5.8654462
#> [13]  -1.0351303 -12.0655224  -3.4354294  -5.0986779  -5.5521331   5.7481401
#> [19]  -6.7256410  -0.4746936 -12.0336272   4.4936678  13.6170631  -3.0623552
#> [25]   3.2308316  -6.3069600   6.4108217  -5.5719022  -2.0959785  -5.8341409
#> [31]   4.1711769 -17.2878638  -4.3763847   5.7037931  -7.5754758 -18.7681513
#> [37]  -4.6373366   3.5694538   4.9244429   8.5550127  -3.8648561   0.6924075
#> [43]   0.8661086   3.0138322   3.5260879  -7.7457080   4.5813975   3.5702790
#> [49]  -2.9222003  -0.5811012

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2686015