Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.99473491 -0.67567432  0.24502000 -1.22333274 -0.82471824  0.32072469
#>  [7] -0.17211648  0.14841789 -1.83611407  1.84855088  1.42912067  0.40395089
#> [13] -0.03017239 -0.24480961  0.58928494  1.91276279 -0.94268696 -2.46231764
#> [19] -1.02544383 -0.24764228 -0.28586822 -1.74848662 -0.25084123 -1.13742413
#> [25]  0.57049362  1.29021357  0.04362872  0.78880318 -0.84056324  0.90515418
#> [31]  0.25449922  1.38220058  0.93854229  0.91786186 -0.94067560  0.83013861
#> [37]  0.89560116 -0.69656643 -2.20593787 -0.40972582  0.38912250 -0.54710839
#> [43]  0.04459710 -1.69797617 -1.71509893  2.32165145  0.45648039 -0.19093917
#> [49]  0.65798308 -1.04941195

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.441596 10.402966 12.526679 12.982638  5.930559 11.366569 15.143268
#>  [8] 15.335369 11.188682  5.302353  8.184412 11.202231 14.920208 13.656063
#> [15] 10.054728 10.498378 10.145410 13.662094 10.200506  4.575942  8.929890
#> [22]  7.023502 18.427911 12.057307 11.451112  9.730045  9.028541 15.383942
#> [29] 13.312331  3.294891  7.120863  8.325457 15.177478 10.444836  6.577768
#> [36] 14.018605  7.546150 14.935894  9.932966  8.622733  3.941629  8.681460
#> [43] 13.474713 12.206589  3.464711  7.829543 12.017955 10.430584 13.987835
#> [50]  7.061852

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   0.79986591  -9.46572120 -10.13879756   6.02973215   1.59673297
#>  [6]   7.97010375   3.70717125  -3.25497141   0.75218053  -0.05948670
#> [11]  -3.55190991   0.21350840  -4.84967859  14.19739157   8.93845166
#> [16]   2.12794255  -0.10681502  -1.08254907   4.10206876   4.62568314
#> [21]  -3.93712048   1.58881776   5.38650550 -14.67024128 -12.06303460
#> [26]   1.12794068 -11.00148325   4.20205420  -1.29959516  -4.60101315
#> [31]   3.25908279  -0.09201283  -7.96209493  -5.92432448  12.80587494
#> [36]  -0.18656348   0.98381580   3.72470282   5.62085122  14.60578265
#> [41]   0.84468035  -5.05428267   3.54297959  10.42766726  12.37597666
#> [46]  -1.25918634  -0.01009637  17.07543799   5.95668392  -8.80685372

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.581384