Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.660047557 -0.802617887 -1.063158882  3.590731485 -0.689418553
#>  [6]  0.643406804  0.282498571 -2.006023767  2.338554367 -0.785143961
#> [11]  0.469947206 -0.006522902  0.305020879  1.942568459  0.307052859
#> [16] -0.522101639  2.415952345 -0.300728979  2.509218556  1.428440206
#> [21] -0.740561832  2.841772919  0.151490884  1.539890962  0.063920201
#> [26] -0.177828857  1.669325016 -0.824371221  0.761445114  1.860905151
#> [31]  0.544394394 -0.916037077 -1.218374250  0.652045420 -1.634129766
#> [36]  0.385493724  0.689538378 -0.993150894  1.760123544 -0.582155596
#> [41] -0.126736116  0.447446087 -0.688241544  1.709868046 -0.430494824
#> [46]  1.573147938 -0.684757179 -1.872461261  2.479413396  0.365751104

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.7360237  4.1757350 16.7817519 12.0137053  3.8633209 13.4121186
#>  [7]  8.1984627 11.5903897  6.3409149  7.3826622  4.6810002 12.8848219
#> [13]  8.7109199  3.2812297 10.1220077  7.9467733  9.3278352 11.4867606
#> [19]  6.2341524  4.4052276 11.2487723  4.9663850 10.5056664  6.2006910
#> [25]  9.9513350 12.6586518 10.6014512 12.5070825  7.4511963  8.2856941
#> [31]  8.6728458  4.6744410 12.1529715  5.3552670 10.7505608 10.0072346
#> [37] 10.0076428 12.0425978  1.9094290  6.8008136 11.6954926 10.1578434
#> [43]  6.2211567 11.3248613  2.9595392 19.1722281  6.2677096 10.9823356
#> [49]  0.9147272  4.9781421

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.1182279  -8.0390238 -11.3847273  23.2571154  -1.8646505  -0.9183635
#>  [7]  -4.0560471 -13.8386747   4.3433503   7.0027527  -9.0376460   7.3045132
#> [13]  -6.8101862  11.7371456  -9.1185588   0.5650185  13.8642661 -16.3011902
#> [19]   3.2036238  -1.4643482  -6.4708932  13.6212986  -3.3747166   5.4573094
#> [25] -14.5784220   6.1415343  -2.7003103 -10.2947819   9.8131353   9.5683960
#> [31]  12.6014207  -9.0443880 -15.4321115  12.8288616  -0.7030961  -3.2243668
#> [37]  -4.7775882 -11.1158102  15.5840851   6.8907659   1.0311752   6.0771989
#> [43] -16.2745091  12.0832217   6.7691581  -7.1193145   2.8300199 -22.0946800
#> [49]  15.3150293  -4.8405705

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2380332