Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.10369816 -0.17738401 -1.45037703 -0.09646903  1.63894000  1.99316732
#>  [7]  0.48136970 -2.14450056 -3.13043208  2.73803908  0.89086673 -1.27097634
#> [13] -3.74353226 -1.01428353  1.68219415  1.77711818 -1.72645983 -2.08621419
#> [19] -1.81377137  0.70794518  1.14790904 -0.70346378 -0.79160586 -3.10326892
#> [25]  2.91029568  1.43978514 -1.57177721 -1.33884932 -1.54140026  2.00726539
#> [31]  1.50851565  0.77396722 -2.23760196 -1.86172791 -2.10534420  2.00122408
#> [37]  0.37620963 -2.33951147 -2.82706768  1.23960273  2.16463438 -0.84463725
#> [43] -1.55956051 -1.66562347  1.32676298  0.50930970  0.01852829 -2.69914761
#> [49] -1.42751021  1.54950333

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.544172 15.630978 10.776547 11.038529  9.516076 12.957501 14.957533
#>  [8] 11.445510  7.051731 11.647893  6.961789 12.835706 10.716986  7.682729
#> [15]  9.215192  9.379261  8.384990 12.825686  7.678717  8.001182 12.677024
#> [22] 11.081293 15.371342  7.607173 18.440317  7.822087 15.779230 10.544672
#> [29]  5.419592 13.893982  6.797799 10.961120 19.517131  8.279565 11.886208
#> [36] 12.846339 14.476786  8.397373  6.567538 12.734416 18.504144 17.377887
#> [43] 10.513425  6.166271 12.919764  9.978018 13.644212 11.682238 12.317741
#> [50]  8.783369

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.4970113   3.4877089   7.8069261  -2.4738197  17.5402579   4.6283707
#>  [7]  -2.0078724   8.1202793  -5.4557910   2.3940700   5.7316027   7.9688759
#> [13]   1.5469277  -3.6419041   5.8268055   2.6278549  14.0079775  12.6399102
#> [19]  -0.1187442   3.7038994   6.2201269  10.5235393 -12.2148347   1.0565024
#> [25]   1.5092098  -3.0631563   1.5738087   7.2677581   0.1090237  -3.8689518
#> [31]   1.2492421   6.9102830  20.8034325  -8.8344623  -9.8227745  10.6763365
#> [37]  16.4371090 -13.4052839 -14.4116692   7.3446830   3.0255626   2.4360963
#> [43]  10.6930961 -15.4452966  -3.9506846  11.1551677   7.9276944  10.1281320
#> [49]   3.1891916  10.5294512

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2275754