Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.27155455 -2.66521739 -0.36730643  2.29574847 -1.08592697  1.34493811
#>  [7] -0.92231999 -0.40456736  2.45224586 -0.35628173  3.02927993 -3.40671056
#> [13] -0.93895167  3.40197037  0.11266977  1.77865258 -2.65737807 -0.84795879
#> [19]  1.00408295  0.52845354  1.71099268 -2.27772775 -0.41129407  0.53537176
#> [25] -0.52901218  1.35694604 -2.72724696 -1.84908216  0.17896067  0.29827631
#> [31]  0.92524645 -2.03076897 -0.48901844  1.05735297  0.56412016  1.78087684
#> [37] -2.06146574 -0.78912379 -1.25899239 -0.08969569  0.69284831 -0.78181033
#> [43] -2.84125978  2.14534699  1.43124397  1.41975667 -1.31782336 -1.35584074
#> [49] -0.85717210  0.90085998

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.304275 14.022087 10.803363 12.854844  5.794000 11.009087 16.250823
#>  [8]  9.589896  9.322512  9.892789 15.399828 12.014535  8.122712 10.839669
#> [15] 14.146307 15.116261  5.431286  7.395177  9.738186  8.727627 15.205701
#> [22] 13.076830 13.734198  9.071738  7.714794 17.934758 13.432696 10.175791
#> [29] 12.989427  7.908535 12.793540 11.692214 13.239848  7.724864  8.714449
#> [36]  9.260386  9.865543 11.076046  7.551373 11.349502 12.461274  9.287373
#> [43] 13.392545 12.601746  3.390579 14.695572 16.007454  8.795006 15.513585
#> [50]  9.067385

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -9.88918103   3.47006705  11.69495432  11.66523486 -24.47244477
#>  [6]  -1.74067540   6.23185567  -0.80662751  10.11897953 -12.64670763
#> [11]  12.39462040 -18.84499434  13.43459074  11.38937456   2.58796905
#> [16]  16.30951189 -10.69803878   4.63816633   3.03599054  -5.15805096
#> [21]  12.39458488  -3.21836768   7.06173052  -4.29223631  -5.31690775
#> [26]   1.57061665   0.04014231  14.40508360   5.75335105 -16.36669002
#> [31]   7.40897844  -7.97155025  -0.97066800   1.72384817   2.94796962
#> [36]   9.74033309  -6.48160853   1.28342950   3.34352055   1.00068015
#> [41]  -3.43290286  -1.53523133  -1.15452991   7.19586575  20.69295326
#> [46] -11.38067331  -8.04169686 -17.27366400   4.39788231   0.37161529

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5422916