Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.90667072 -0.42374889 -0.05228525  0.12772481  1.32699636 -0.35600194
#>  [7] -0.90423067 -1.29017314 -1.40459497  0.91076240 -0.63884260 -1.86645805
#> [13]  0.12960368  0.10516009 -0.74894127 -0.68369426 -1.28364661  0.53289604
#> [19]  1.52555445 -0.55004263 -2.64184442  1.34392480 -0.15021985  0.65232110
#> [25] -0.77182743 -2.52797797 -1.59432698 -0.22875721 -1.07903216 -2.04791726
#> [31] -0.37141048  0.53490431  0.16088665 -0.18725172  0.92675545 -1.41207478
#> [37]  0.51238554 -1.87556982 -0.29950982 -1.18318550 -2.33603465 -2.01412006
#> [43] -1.44725536 -0.81292811 -0.18512339 -1.39315335 -1.91633240 -0.87095085
#> [49]  0.12711293 -0.89868361

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.171849  6.485314 10.971046  4.507654 15.093525 10.415510 10.273590
#>  [8]  8.280390 11.390665  8.502828 11.029926  7.823303  7.214761  4.170893
#> [15] 14.440635 10.681627  7.666915  9.770974  2.621480 12.357806  8.938603
#> [22]  5.513752  7.648880  8.504061 10.461674  7.763976  9.251072 13.752187
#> [29]  4.608310 11.008069  7.020059  5.034448 12.448844  8.286931 12.974496
#> [36] 11.081021  3.076542  9.369079  6.329282 10.578140 12.406345 10.295487
#> [43]  4.458202  3.555055  8.499555 13.489742 10.265959 11.648443  5.510570
#> [50] 13.448285

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -2.11903618   2.38718523  -0.06894992  17.94618043  -5.48899556
#>  [6]  -4.82710747 -17.53418938 -18.39809318  11.89275971  -0.22098036
#> [11]  -1.93661363  -2.58239478 -12.35587927  14.93605585 -15.41696381
#> [16]   8.78409692 -13.65029473  -1.14160850  -6.41656028   6.00270458
#> [21]  -4.59611647  -0.29255459  -4.63513178   3.77252706   1.67352100
#> [26] -15.65333034 -10.22615875   1.30144133  17.46944937  -8.16846674
#> [31]  -7.83236903   9.30378519  -1.13844507   4.18915259   0.60449953
#> [36] -21.51520313  -6.13641807   0.23005857   7.81126626  11.25453660
#> [41] -20.81361314   4.69646085  14.69408059   2.20761248 -18.44175264
#> [46]   0.10646557  -5.02643522 -22.21731975   0.91288533  -5.03591602

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.396024