Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.09174687  1.79767205  1.28003874  1.14716899  1.55804332 -1.56142885
#>  [7]  1.65184288 -0.18900264 -1.57765407  0.11861931  0.69029020  1.37584215
#> [13]  0.44294846  0.87491117 -0.52727834  0.35704320  3.72697923 -0.07030235
#> [19]  0.77242165 -0.77280561  0.21024284 -0.76179639  1.16310866  0.75678367
#> [25]  0.43724744 -0.31560027  0.85991550  1.23332047 -0.70745478  0.63431421
#> [31] -2.12217726  3.17151997  0.94799458  0.94719323  0.11241348 -1.90624800
#> [37]  1.48444855  0.56052947  0.79574088 -0.56411294 -2.03382171  0.89800672
#> [43]  2.56026220 -0.18041184 -0.09725541 -0.85400974  1.67961465 -0.97078593
#> [49] -1.75383576 -1.14151276

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.2698131 10.9184636  9.1340754  3.2686247 12.7647743 10.9558439
#>  [7] 10.2971568  7.5512584  4.4632906  6.6133518  7.4305508 12.3929695
#> [13]  8.8183550  7.8122768 11.1327091  9.4427339 14.4738020 15.0422590
#> [19]  5.6345405 12.0090436 10.6125393 11.3758951 11.7722559 -0.7483179
#> [25]  7.5864008 12.3640133 18.0283839 14.5958312  5.3531121 10.4557457
#> [31]  5.7896041 12.4088267 12.3289630  1.2739909  8.0276136  8.6312094
#> [37] 14.5906165 10.8544110  6.8332973 11.9338608  8.7803038  9.0577594
#> [43] 12.3143103 10.4940227  5.4232243 10.1005299 14.3033122 12.4370948
#> [49]  5.0622690  9.0679993

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -4.8436342   9.8606105  -3.1841004  -8.1758675  11.1054240 -14.7079115
#>  [7]  -0.9282358  -7.1537925 -22.6399590   1.3037784  -3.5366294   2.2782667
#> [13]   3.7971176 -12.7565899   2.3816100   5.9997307   5.7648250   2.1831512
#> [19]   1.9435930   2.8587803  20.4642375   7.7759919 -13.4930595 -15.0974308
#> [25]   7.0899000  10.6864469  10.5596527  15.2284968 -12.7752032  12.9426320
#> [31]  -9.4866611  -3.4341240   2.0159903 -14.6498235   2.0035168  -1.5120684
#> [37]   0.4634403  -9.1669235 -13.1891704   4.6585015 -13.5983730  -6.4659582
#> [43]  20.3333693  -3.3645195  12.9156544   0.3390967  18.7264382 -10.9665054
#> [49]  -1.5280778   0.9118081

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5346439