Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.36902914 -3.52517891 -0.49418611  0.76666072  1.14448755  3.60789114
#>  [7] -2.12204962  0.42313706  1.33614976 -0.17972039  1.57342820 -1.61995940
#> [13] -1.66050178  2.14414368  1.25663648  0.92141008 -1.64654623 -0.17114832
#> [19]  0.64738990  1.04405041  2.36098043 -1.90840394 -0.16412058 -0.59818762
#> [25]  0.86907757  2.45594635 -2.34174828  0.04013271  1.31135174  3.46284802
#> [31]  0.69644334 -2.28109465 -0.05274222  0.84435545  0.18863948  1.59186057
#> [37] -2.29870211 -2.99106727  1.48003151  1.95383472  1.97094960 -1.14471893
#> [43] -1.23558464  0.34139262 -0.14496177  2.46696883 -1.92339397  2.08865493
#> [49]  0.12059936  1.68745716

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.358474  5.779429 14.116789 13.800592 10.211033  5.077487  2.338040
#>  [8] 12.365934 15.148946 12.999157 15.660495  4.557734  9.725063 11.328508
#> [15] 15.453820  9.797451  8.549457 14.027304  9.885171  8.968467  7.051280
#> [22]  7.566975 10.076106 10.784038 13.442950 11.717046  9.395437 16.207838
#> [29] 10.594656 13.817398  7.691547  5.654200 10.914966 12.725753 14.776485
#> [36]  9.343146  7.879121 16.000130  8.407346 11.601521 11.421350  9.209637
#> [43]  7.780531 12.514704 14.665312 12.083080  1.888731 12.309849 11.577344
#> [50] 12.208580

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.5471148  -4.9994569  -7.2798541  -0.9912684 -10.1597263  -2.8033955
#>  [7]  18.9652503  -6.5914680 -10.1816931  -0.1781617   6.0008373   5.8658650
#> [13]  11.4872241  15.6282667  -1.9810669  12.1533307  11.0177160   8.9180887
#> [19]   3.1015867   5.8839815   9.3829799  -2.9463147   6.1787947  -4.3044019
#> [25]   2.5967585   8.4087325  -4.7965569  12.4261428   0.7683657 -11.6670855
#> [31]  -3.8946870   5.5601253   4.3665003   6.1113266   5.6593606  10.1259571
#> [37]  17.3975014  -8.0185360 -10.0351824  -3.2998530  13.4438475   0.4382631
#> [43]  -2.4611856  -1.9025477   1.0814615  10.0953496  -1.7324126  -0.7280944
#> [49]  -7.6199142  -5.0827458

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.08162184