Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.595910737  0.267191999  0.848569947 -0.507447574 -0.646034917
#>  [6] -4.257850499 -0.035694373  0.078345246 -0.834036308 -1.575943236
#> [11]  0.185234509 -0.388365324  0.313261581 -0.008949984 -2.026155688
#> [16] -1.200067629  0.413582232  0.509159940  1.045065000 -0.004963429
#> [21] -2.593852446  0.340312063 -0.599362150  0.537487903 -0.097788736
#> [26]  1.557726531  0.389811204 -0.710824460 -0.340781739 -1.503134591
#> [31] -2.419590875  1.182822537  0.366961711 -0.450959359  1.692674799
#> [36]  0.499478317  0.784691326  1.080236728  0.099177244 -0.965163589
#> [41]  0.189385982 -0.013039840 -1.396207545 -0.098307935 -0.691884000
#> [46]  0.259814276 -0.660681248  1.589159864 -1.254688292 -1.027730914

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  4.6848482 10.8713659 11.7303398  8.1869568 14.5088328  6.1894809
#>  [7] 12.5874452  9.1801214  9.8036004 14.3343454 10.0391580  7.0369040
#> [13]  2.9831927 14.6240372  9.3389703  5.0298326  7.6664602  7.5541144
#> [19]  4.8088658  5.5680250  3.9795326  4.3395025 10.6718476  5.4363980
#> [25] 18.5761786  8.1369254  9.3613227  6.9151198 13.0004528  9.8149215
#> [31]  6.7486872  4.4771533  3.4905516  5.7193919 10.2872701 -0.9627388
#> [37] 12.5034124  6.4785064 14.4523442  7.1505491 13.0334406 11.2923714
#> [43]  7.8601842  8.4470371 12.4144935  2.6173331 10.5416616  6.0060012
#> [49]  4.4394549 10.9490616

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -4.2367411  12.8267983   6.8216344  -3.9925562   0.9546292   4.5536593
#>  [7]  10.3927819  -4.6367709   6.6386144 -15.7378306  17.5095770 -16.9457932
#> [13]  -4.2509082   5.8731028   1.8507450   7.5020921  14.2952401   5.1601459
#> [19]  -0.5793813  -7.4337777   8.3897898   9.2512317 -13.6676280 -16.4992655
#> [25] -12.2780245   0.2379043  15.0180317 -10.4840370 -10.5566038  -8.9389075
#> [31]  -2.6841414  -7.2517167   8.0601903  -7.0702398   5.6592195  -3.1389439
#> [37]  -0.1202227   2.1272410  14.8027348   0.8196387   1.0384784   0.4288503
#> [43]  -7.2089653  12.3322179   7.9749102   1.0166767  -1.5272825   5.7933724
#> [49]  15.7413621  -2.8373839

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3665101