Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.073682169 -0.963141503  2.271423247 -1.890651759 -1.363433217
#>  [6] -2.410256687  0.161225151  3.265013827 -0.279200979 -1.481732381
#> [11]  0.150476149 -0.114820006  2.243680335 -1.181853265 -0.612217762
#> [16] -1.029229353 -0.772805750  2.575714334  1.470541397 -0.527958313
#> [21] -0.132460434  0.407426975  1.501230545 -2.560405665 -1.301361740
#> [26] -1.844734773 -1.050559598  1.903732136 -1.205940400  0.537772740
#> [31] -1.540444773 -1.018183178  1.150248985 -0.375512786  1.191196329
#> [36] -0.354306338 -1.097658521  3.361829514 -1.237245637 -0.008153511
#> [41] -0.448359968 -0.708969545  2.580087590 -2.597231013  0.533249556
#> [46] -1.569932497 -0.706750295  2.656718421 -0.225550134  1.955809784

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.829698 14.122080  4.141106 12.611150 12.122925  9.753691  9.832956
#>  [8] 10.407455 15.610038  7.140882  4.936264 12.393622  7.374563 10.204362
#> [15]  9.823768 11.200525 13.611469 11.324601 11.349955 10.284257 13.306449
#> [22] 11.001408  9.198947 12.785562 17.180114 11.696883 12.894964 11.694855
#> [29]  8.955081 10.471858  6.889018 17.622269 10.421375  8.622904 12.037359
#> [36] 12.384476  6.225073  8.697046 11.542084 10.681301  2.003749 15.795751
#> [43] 10.426947  8.038178 12.926232  9.153165 11.580944 11.105878 14.027011
#> [50] 10.197251

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.95933133   1.20107273  -5.25020007   7.30833043  -4.46369119
#>  [6]  -4.73186268  17.09419281  -6.99583436 -13.45716336  13.63430067
#> [11]   2.94472419  21.15058645  -9.34196424  -9.89891775  -3.04333161
#> [16]  -1.13041522  15.76304708   7.10508898  -1.33203633  10.00547517
#> [21]  -2.96344519  -4.55232056  -4.34340919  -6.91482384   5.64171567
#> [26]  -7.23643892   7.96123389  -0.26950647  10.88305669  -3.83516799
#> [31] -19.00820740 -10.64746555 -14.31503139   0.70119745   1.29426403
#> [36]   6.72319602  11.49076028   3.44878699  -3.09682947  -3.77360041
#> [41]  -6.02552215  -1.02550421  -5.98502454 -14.62759812  -6.92589397
#> [46]   0.19026194   0.05896725  -1.65669123  -2.80350469   2.95176821

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4531591