Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.99163959 -1.77907804 -2.24542537 -2.40741196  0.40350767 -0.99571684
#>  [7]  0.52121144 -2.45120855 -0.79108500  0.27677000  0.30775178 -1.53070897
#> [13] -2.52860248 -1.37803830  0.29844931  0.04798858 -3.81509228 -1.06399423
#> [19]  0.67847733 -0.01718847  1.26859346 -0.62350707 -1.87942318 -0.62442602
#> [25]  0.57286067 -0.95659486  0.13814689 -0.58058514 -1.19363045  0.95788892
#> [31] -0.06265274 -0.91623025 -0.84074588 -2.23514864  1.60597140  1.24769531
#> [37] -0.72697763 -3.07494648  0.58077465 -0.81468599 -0.82880182 -0.63889153
#> [43] -2.32019857 -1.76664791 -0.56972800  1.23021812 -0.71802624 -1.05369583
#> [49] -1.94042688  0.11974706

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.312880  8.581566  8.009613 12.248423  8.164927 15.735798  9.956354
#>  [8]  9.118342 11.126719  5.881165  5.477265  7.668773  6.088452  6.421858
#> [15]  2.183765  5.592283  5.194619  7.627088  6.703774  8.657182 12.597261
#> [22]  8.991118  9.684923  5.430874  6.647672  5.670072  2.833933  4.878329
#> [29]  8.264543  3.780958  6.896553 10.018743  8.541537  4.180824  8.789704
#> [36]  7.211556 11.779274  4.640712 11.790993  8.520958  3.564319 11.342946
#> [43]  9.639777 13.103542  5.745550  7.360819 10.560289  7.696809  6.343594
#> [50]  6.340295

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  14.8626261   5.0120047 -13.2040561 -11.4754626  -1.8762598   6.6203266
#>  [7]   1.7415938  -8.1853577  16.5188943  -3.0439204   1.1286059  -5.0784312
#> [13] -16.4465865  -5.6886658  -7.7305494  -3.7442403   0.7551978  -1.4979043
#> [19] -11.0898303 -11.4805857   0.8130515 -10.2994783  10.9310800  -7.6742858
#> [25]   0.5219829   4.2069322  -7.4879910   9.6334244 -10.9520707 -13.3037145
#> [31]   3.3653507   1.5430012  -6.9439758 -11.7003026  -8.8437016  10.5258149
#> [37]  -0.5879733  -3.1881688   9.4552369  11.9390953  13.3169085   8.9460150
#> [43]  -7.9085979  -5.4713031  -9.7792102   2.5800385   2.7207549   2.2283116
#> [49]  -6.5547890  -3.2008821

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5712725