Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.8578030 -1.0872705 -2.5052587  0.3604719 -1.2557156 -1.2898563
#>  [7]  1.3015655 -1.4584015 -0.6199409  0.9143494 -1.5965224  1.5655413
#> [13] -2.0417820 -1.2342938 -0.6266842 -1.7350127  1.4172888 -0.2128321
#> [19] -0.5429862 -1.6961127 -2.2247394  1.9535473 -2.3112297  0.3174284
#> [25]  0.1048537 -1.4878209  1.5880652 -1.2407093  0.3143003  0.5780561
#> [31] -0.0891347  1.3375657 -2.2738781  1.2863029 -1.3397477 -1.7352367
#> [37]  0.9210828 -0.5261649 -1.0434076  0.2522105 -2.4493895  2.1214304
#> [43] -1.7745353  1.1453805  0.2433633 -1.0139414 -0.5844183 -0.5691995
#> [49]  0.8908136 -2.1406243

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 16.6144323  8.8708479 14.7204449 10.1196709 10.8882256  9.0026603
#>  [7] 10.4159994 12.0950269  8.8097809 13.6348982  7.3803877  2.1348197
#> [13] 12.5778715  4.8256702 13.3289552 11.6656756  4.8128604 10.1600866
#> [19]  8.7820657 10.8449088  5.6298565  5.5672016 13.8346370 14.5211734
#> [25] 12.0823212  9.9232040 -0.1384961 10.8117751 14.4883110 11.5340565
#> [31] 13.6293833 11.6328109  8.5207578 12.0787324  7.4941334 10.7286267
#> [37]  5.1345236  7.3391837  6.1197758  8.1617187  8.0871928  4.8286609
#> [43] 12.8876855 12.4738486  8.5218788 11.8654019  7.2142282  8.7604304
#> [49] 11.5731707  5.6149671

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.30553053 -10.64796472  -9.57633480  -3.82135451   2.63014481
#>  [6] -18.13613512  -5.17124249 -16.90817388  -7.10974393   3.00056361
#> [11]  -1.69028312  11.73187655  -7.06878241   2.67484788  16.49418824
#> [16]  10.10819395   8.91105158 -22.41695940  11.97876692   9.16277269
#> [21]   7.14276479  -0.27510934  -0.08855464  -7.02855040  -5.11721349
#> [26]  -3.99705352   8.98343084  -2.08234778  -8.93003028   4.01322087
#> [31]  14.90726999   6.56462148  -0.79438809  -9.93469572   6.74314062
#> [36]  -2.20127183   2.04764458 -18.70359877  -4.06716782   5.75597765
#> [41]  -9.38432833  -7.63835396  -0.91910555   3.04639475  15.53087550
#> [46]  -1.99397079 -11.94597244 -10.50857015  -0.31776509  18.70129423

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.403799