Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.71892657  0.41162406 -1.87636789 -1.08965427 -0.60453194  0.31711732
#>  [7]  1.84996684  0.42061284  0.02765663  0.23863713  0.47220110  0.53062202
#> [13] -3.52600263 -0.91175187 -0.52816315 -0.02273866  0.78509364 -2.66025075
#> [19] -2.49772973  0.66786500 -0.23101142  0.52933218 -0.92084904  0.78943435
#> [25]  1.34755096  0.28299320  1.12026667 -0.67496086 -1.01707292 -0.56609947
#> [31]  0.79220215  1.31313035 -1.22438441  0.85784235 -0.19980021  0.80420058
#> [37]  1.76224180 -1.85069951 -0.76038526  0.79650378  0.34195318  1.50394772
#> [43] -1.25289695 -2.00493764  0.02235325 -0.92761582  0.35480787  0.21833480
#> [49] -0.31470805  0.86136800

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.833783  7.043119 15.609146 12.050840 10.105321 14.415504  4.182824
#>  [8]  7.789091 14.380744  9.721821 12.552859  8.300981  6.425688  7.303231
#> [15] 11.993676  8.479694  3.560105  8.827279 12.728376  8.406410 11.812902
#> [22]  2.104756 12.323921  8.267278  6.425057  9.302092 10.012078 12.289166
#> [29]  6.740077 12.066882 14.511912  2.599093 10.440178 11.734009 12.210360
#> [36] 11.474773  5.872568  7.476790  8.341811 13.522894 12.496269  7.254937
#> [43] 13.322933  5.111789  8.886959 13.318213 10.153198 11.501412 12.689383
#> [50] 10.721439

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   5.0298847  -8.5373448 -13.4427870 -11.3618078  -1.8922318   8.9150713
#>  [7]  -1.7769627  -4.0790500  -7.5043552 -23.3532331   3.9861687  -6.5389373
#> [13]  10.1869596  12.5512812  14.5805059   1.8522504  -2.7174981   7.7368428
#> [19] -11.9551328   7.0755905  -1.7443145  -7.7531352   5.5320824  -5.6804841
#> [25]   0.8826933   8.4431874  15.7732542   2.4168876 -23.2547037   8.3219376
#> [31]  12.9779228  -1.7653747  11.9168935  -4.0728572 -21.4746769   3.5626728
#> [37] -13.2658621  -1.1562120   3.1214021   6.8492797  -9.6543654   2.6525653
#> [43]  -5.3058839  -7.0757678   3.4937867  -1.1887341   7.4135665   3.1625296
#> [49] -10.8790144   9.7993029

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2792163