Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -2.92981717 -0.31822380 -0.45343423 -1.40085223 -1.02226175 -0.45902073
#>  [7]  0.54964126  1.61550793 -0.06634703  1.37803855 -1.34608779 -1.33785320
#> [13]  1.47002481 -1.28656530  2.65992210 -1.72299088 -0.17951040 -0.21724079
#> [19]  0.08689088 -0.06333984 -0.24658464 -0.18323378 -0.09510946  1.04574924
#> [25]  0.88604860 -1.14979797 -0.34210184  0.52348290  0.14316540  1.29548910
#> [31] -1.47744319 -2.49684097  1.40728273 -1.38646163  0.51309891  1.12493212
#> [37] -0.12071823  2.75362130 -1.54027522  1.49174923 -1.65522663  0.32217653
#> [43] -0.39672385 -0.32460098 -1.02144563 -1.61803875 -1.03907254  1.02943513
#> [49] -0.73381389  0.94454805

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.881850 13.102724 10.752018  9.435225  5.091120 11.747270 12.918228
#>  [8] 11.104040  3.549607  8.102399  9.598557 13.448612  5.190279 11.571462
#> [15]  9.485235  8.709661  9.096281  4.253023  5.005171  9.402083 12.306531
#> [22] 12.975396 14.801924  3.244975  7.192526  9.809532 15.832406  7.858283
#> [29]  9.021425  5.840380 21.183010  7.639321 16.904630  5.599250 11.590004
#> [36]  9.909869 11.323862 13.664189 11.177557  9.322112 17.339590  4.493372
#> [43]  9.389369  8.895004  8.711852 11.270655 10.461864 12.856380 14.225174
#> [50]  9.840556

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -9.8028425  -4.2086012   9.2956492  -6.3620293   6.5542688   6.9456617
#>  [7]  -4.4044688   4.0025695   5.3236894  -0.8584609   1.2432906   9.3521397
#> [13]   4.4314959  10.0714714   7.5571647  -5.0714555  -2.6832362  -4.1554973
#> [19] -21.0447522   2.2451645  -4.3359588 -12.7639766 -11.9058734  -2.0702307
#> [25]  -4.9544250  14.0121009   5.4584065 -22.6293631  -6.9559593   1.4967755
#> [31]   4.8285004  13.9941351   0.6538003   7.2594192  -5.5740558  -7.2938388
#> [37]   3.0131879   0.9450216   6.8686882  -0.2459526  -4.0849575  -2.2648382
#> [43]  12.9910338   9.6041054   7.3303647  16.6177964  15.1415035   9.9859188
#> [49]   4.9452122   9.8049713

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3910858