Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.93720958 -0.18803905  0.06575471 -1.87189945 -0.05909928  1.36235526
#>  [7]  0.87027887  1.41625083 -0.48723072  0.29387193 -1.18022996 -0.62850006
#> [13]  0.73660422 -2.24464488  0.77847746  1.40405171 -1.69775354 -1.41273504
#> [19] -2.24952550  1.17966857 -0.90417872  1.56377677  2.05744052  0.20763055
#> [25]  1.14102786 -1.25285448 -0.94334207  2.01431675 -3.22673720  1.21488582
#> [31] -1.12949051  0.11554341 -0.13371489 -1.14702038 -0.61593840 -0.45455667
#> [37]  0.98892692  1.08360218 -2.09307070  0.76031989 -1.38824014 -0.02382626
#> [43]  0.77190147 -2.28298479 -1.55471071 -0.48576669  0.65517904  2.70613970
#> [49] -3.14533598  1.28325987

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  3.3154297 12.2145473  7.1633664 16.8236097  8.6848794 10.2156310
#>  [7] 13.2848188 14.4266414 11.9716525  8.2610692 -0.8732159  7.2455102
#> [13] 10.9909884 20.6122921  1.8488003  7.0495613  8.9576352 11.2454740
#> [19] 13.7260475  9.9978464  4.3000738  7.0076402  8.2453039 13.5624819
#> [25]  4.8216376  5.1379994  6.9248592  6.6680339 10.7483001 10.4365136
#> [31]  4.1997140 10.8192265  9.0808934 14.2768156 10.2293462  8.6989882
#> [37] 14.2524178  9.1524327 13.0859385 11.1644922  8.0272972 18.3609315
#> [43]  7.3215434 14.7674030  7.4377295  4.9245855 15.0741743 12.2578088
#> [49] 14.0984361 14.9944266

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -6.8318311   5.9714177   5.9002565  -0.2376303   3.3081385   1.2717711
#>  [7]   2.1598215   9.3954741  -5.8126050  -1.9411837  -6.5125137  -4.2059913
#> [13]  -6.9643258 -12.2046855   3.4849359   7.9779198  -3.5875995  -1.9921327
#> [19] -12.1262481  -0.7942393   0.1379527   7.6800937  -7.1186260   0.5478052
#> [25]  16.1236854  -7.4616158  -0.2192609  -2.2547155  -5.9503647  -2.4328655
#> [31]  -3.9496905   0.2205591   3.3658967  -6.7425053   1.3348765   8.3915656
#> [37]  -5.2666381  -0.1579746   7.8323659  -4.8352198  -3.3770821   3.2011094
#> [43]   3.7445419  -1.1272870   2.1131732  -4.4308754  11.2171852 -11.7780385
#> [49] -13.0079878   4.5744435

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5795755