Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.3042535 -1.1413830  1.1711161 -0.9079476  1.5326945  0.9264625
#>  [7] -2.5701007 -2.2592014 -1.1135071  2.3349786  0.9572472 -0.6904751
#> [13]  0.5963440  0.4995287  2.9414687 -0.1254194 -0.6274214 -0.2919200
#> [19] -1.3163579  2.9842174  0.7068549 -2.7244619  1.9618038 -1.5940081
#> [25]  1.7133185 -0.3086932 -0.3250860 -0.5301716  0.5237849  1.4146075
#> [31] -1.1808968  0.1890727  2.5258995 -1.7684876  2.2985247  1.9341919
#> [37] -1.7813220  0.2826558  0.5191287  1.7417637  1.8917123  0.4240693
#> [43] -0.7284536 -0.4639850  0.7980082 -0.4705474  0.5218003  0.1565464
#> [49] -1.9762182  2.7913027

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.301598  7.091509 12.105007  9.179384  3.465990  4.033371 13.091831
#>  [8] 15.590804  5.825882  7.397472  5.319112  7.037653 10.541359  9.454498
#> [15] 15.457816 10.433689 15.365969 13.415555  9.758743  6.109825 11.576022
#> [22] 12.067461 18.053825  8.143267  8.960171  9.350946 11.485283 13.433478
#> [29] 10.678219 10.135465 10.330150 12.266758 12.219689  9.230670  4.735432
#> [36]  8.927459  8.901356 15.034027 11.399177  7.062670 11.078711 17.399001
#> [43]  9.966291  8.936859  5.865866  5.532560 16.365115  9.481456  8.538296
#> [50]  8.842340

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   4.5142612   8.2916842 -12.2118399   2.7099676 -21.3918734   1.8271053
#>  [7]   5.0234967 -10.1303549 -13.7857876  -5.7032990  -1.5416756   7.3502304
#> [13]  -1.2149312   1.8901830 -11.0725634   9.9529843  -7.6687509  -1.1335402
#> [19]  -6.9322025 -11.9872733  -6.8824783   4.6317821  -3.1845946  -6.4911237
#> [25]  -7.6627871   4.2291655  13.0600385 -10.0732661  -9.8967320   1.3244871
#> [31]  -3.2757851   5.9330646 -20.6076736  -3.4264157  -2.2237233   0.5856936
#> [37]   7.8547436  -8.3921995 -12.1509197  14.6573239  21.5304538   4.8492649
#> [43] -14.3076622  11.8290996   1.1883572  -2.2100836  -2.4940064 -21.7252664
#> [49]  -5.1399571   0.9764098

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.06258114