Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.61398470  1.52913374 -0.68402908  0.95568294 -0.02415095  0.97878511
#>  [7]  0.36444032 -1.36001574 -0.26187095  0.01517219  1.16138252  0.48156382
#> [13] -2.89940688 -0.13295347  1.23112381  0.80727957  0.40156802 -2.40039321
#> [19]  0.08153539  1.98902722  1.29566283 -1.51550334 -1.26041899 -1.28380786
#> [25]  0.77246203  0.38274525 -0.58344888 -1.43261854  0.66688037  0.71855472
#> [31]  3.31593340  1.88054234 -1.93821969  1.20857938  1.70670290  1.51579742
#> [37]  1.53387853 -1.46310373  0.38030576  0.82675126 -0.84160643 -0.73409790
#> [43] -0.79837238  1.16011986  1.21816187  1.57287871  1.88359596 -1.07115016
#> [49]  1.15003549  0.25920363

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.404560  2.210941 13.231609  6.720004  9.766494 14.875906 14.436406
#>  [8] 11.771294 11.436338 12.735922  5.517380  7.387984 11.520509  9.165524
#> [15] 14.379806  8.582598  9.161939 15.574262 13.146866  8.846740 10.136341
#> [22]  4.936781  9.271121  9.318560 11.731215 12.188512  2.483714  7.859087
#> [29]  5.028583 15.689331 14.901189  2.051028 12.079667 11.052659  9.304683
#> [36] 12.324668  8.430176 11.378105  3.938852  8.813106 13.186594  6.415490
#> [43] 12.553863 14.116317  5.368424  9.368923  6.530129 11.306925  8.110700
#> [50]  9.749729

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -12.8025150  26.9029226   9.5438642 -14.9255998  24.8963736   1.0579021
#>  [7]  -5.2745574   6.6548523   2.4003224  11.4558972  19.2394531  23.7598809
#> [13]  -6.6142869  -3.4058689  29.5229417   9.6201135   0.1388801  -7.3726199
#> [19] -14.7501921  17.9511822   6.1767219   5.2133035  -6.7623106   3.9675564
#> [25]  10.4407118   1.0828788  -0.7022796  12.9418495  -1.4218403  10.7993556
#> [31]   4.0402650   0.3196767  -1.1641857  -9.7892071  16.7929161   8.0492188
#> [37]   0.5838712   5.6024930  -2.5107193   5.8856457   0.3697532   0.5754394
#> [43]  10.3318164  -5.5130042  11.1228488   9.7410795   4.2962326   6.1820260
#> [49] -14.7141685  -5.2708439

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4560466