Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.469224322 -0.182656080  1.051671806 -0.419593324  1.249812371
#>  [6] -1.032923358  0.580887933  0.902558981 -0.417216606  2.514706566
#> [11] -2.631211136  1.209454769  0.666647230 -1.859939449  2.996775482
#> [16] -0.302442765  0.082514327  1.188188892  0.078046498  1.162502598
#> [21] -0.380546119 -0.024627554  0.756589410 -1.204664082  3.186521086
#> [26]  0.857458800  0.739330557  1.931348957  0.006777637  2.135588733
#> [31]  0.181595053  1.209244332 -0.526779530  0.608563099  2.900052311
#> [36]  0.855047494  0.738917146  1.584470994  0.540770943  1.026605919
#> [41] -0.360277020  0.984746870  1.336689229 -1.868517559  3.674309134
#> [46]  0.584122248  0.027006676  1.054199623 -0.921504237  4.098685880

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.614056 15.528431  9.324785 11.878628  9.831283 10.599162  8.049248
#>  [8] 15.122481  8.726876 11.558353 12.762316 11.009494 13.783911 11.382927
#> [15] 12.559742 11.623435  7.418047 12.336158 13.892911  9.511910  7.953048
#> [22] 14.895182 14.438772 12.005894 12.828966  7.241560 11.342321 10.353790
#> [29] 11.099995 12.985391 10.331293 13.798331  5.466114  9.681875  9.114718
#> [36] 10.303538 11.442502  6.152629 18.635250 12.331487 11.281092  9.617728
#> [43] 14.177764 11.848149 15.086582  7.737664 12.322478 10.172744 17.884831
#> [50] 10.591329

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.5531422 -14.4646598   6.1433431  10.9082007  11.0846006  -2.5740468
#>  [7]   2.9401702   4.4048501  11.6715829  -2.7860121  -1.5981711  -0.4327956
#> [13]  25.6861559  12.6505972   4.1498473   2.9006580   5.0298609  -6.5051297
#> [19]   0.8148780   3.2619771  16.7752091  -1.2859024   0.7263778  12.8220730
#> [25]   8.6511565  -1.5083912  -4.3936747   2.8739756  -2.3451777   4.2209173
#> [31]   1.1726633  -9.1947482  -2.4530430   9.3588797  -8.5295597   8.8359977
#> [37]  10.5318501  -0.5609048  15.1295139  10.6291481 -14.1151963  13.2931309
#> [43]   9.0115473  13.7396236  10.2389682  17.2386420   1.6534411  16.1752049
#> [49]  -1.0326052   8.8521292

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3450747