Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.01632429  0.46828372 -0.26889095  0.89099239 -2.61453646 -0.31539215
#>  [7]  0.55499109  0.74917560  1.01613344  0.61670560 -0.02533263  0.39558477
#> [13]  0.91948409 -0.66962515 -1.08234776  0.63352520  0.14391414  0.51158184
#> [19] -0.02461433  1.05379562  0.81370402 -1.11163750  2.45720152 -1.28758239
#> [25]  0.21556760  0.85812365  1.03317048 -0.06518007 -0.03346701 -0.99979261
#> [31]  1.14726121  0.67232823  2.05850205  0.14011550  0.64118184 -0.29893836
#> [37] -0.13826335 -0.03011759 -2.95596209 -0.71938711  0.34536077 -0.08607119
#> [43]  0.82750063  1.09636798 -0.27967194 -0.55069031  0.63979154  0.80718562
#> [49] -1.93554377 -0.59028049

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  5.890463 11.185910  6.619029  9.722372 12.563946  7.301442 18.064782
#>  [8]  7.264608  9.990981  8.231424  7.112265 12.381209 11.956641  9.754718
#> [15]  6.318616  2.045894  9.062766  6.656764 14.036126 10.139258 12.478045
#> [22] 12.224428 12.641219  4.232093  7.250504 11.432208 15.793708 10.290937
#> [29] 16.916883  9.246182 10.195515  9.975529 10.266941 15.127001 14.384134
#> [36]  7.931162 10.845969 11.950879 12.345416  5.408925  4.784265  7.959936
#> [43] 10.254656 15.252485  9.261935  8.060202  9.391697 10.232644 13.939989
#> [50]  6.150765

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -6.182484381   6.630156987   3.169827722  -4.019132952   7.633284069
#>  [6] -10.494172541  -4.825321660   2.771447901 -15.824650998   6.170545607
#> [11]  -7.224130823   2.692399960  11.771783209 -21.488603065   6.275173681
#> [16] -13.114553748  -5.784623649  -2.213500987   6.797359913   2.494852373
#> [21]  -7.125415984 -13.937573218   6.121011749   0.453268585  -6.055238410
#> [26] -10.136902968   2.043130257   4.285404279  -8.863983218   3.108353296
#> [31] -17.906559857   5.237647531  11.964577471   0.005993301   6.831089472
#> [36] -10.619882755  -2.340211104  -4.267901584 -17.704636006   0.521949785
#> [41]  -1.984691218  -0.488898024   6.844181481  -3.497757351   1.088402027
#> [46]  -3.150943663   0.555274527  -5.102108644 -12.562226102   9.585029798

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6512571