Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.049508589  1.492039206 -0.088459134  1.107340434  3.730799756
#>  [6] -1.033313241 -1.906486616  1.050351960  1.277192604  2.691323076
#> [11] -0.799000013  1.335812490  0.580347103 -0.664468247  4.084073072
#> [16] -1.796347551  0.020186412 -0.626698147  3.321790837  4.662792096
#> [21]  0.676443698  1.959219389 -0.271853751  2.467518656  3.977697148
#> [26] -1.277121786 -0.009493573 -0.410292110  1.397873288  2.215093886
#> [31] -2.029864131  0.101626393 -0.274300334  0.166346156  1.635097818
#> [36] -0.131212015  0.779690589 -1.940941198  0.336577886  1.327634784
#> [41] -1.495881687  0.231063504 -0.011847359  0.119362733  2.426198545
#> [46]  0.352249662  0.039665903 -0.689787869 -0.166387801  2.662726247

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.03092099  8.51029984 11.31844722  9.54463767  8.74721383 12.19945167
#>  [7]  8.51557160 12.66962770  4.97591076  6.99932960  7.80851755 14.01599121
#> [13] 15.28708194  6.23769228  0.08438752 11.93560677 12.06793789 10.96554328
#> [19] 13.00082950  8.62746678 14.11446103 14.23940851 17.38437572  6.55948262
#> [25]  8.62905819 15.49856191  5.39186685 13.13259891  7.09799965  6.55714447
#> [31]  6.81278314 14.22496137 12.81009232 15.45042962 13.22349162 15.17393637
#> [37] 14.57038557  7.76986671  9.11490998  9.99510388  9.86736802 11.68180707
#> [43]  5.09289585  4.68136648 10.40355690 12.27102095 10.97794036 16.04818356
#> [49]  9.03644985  6.25242422

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -2.7837847  -1.0747941  -5.1371784  -8.5050632   0.4846082  -5.6389936
#>  [7]   2.1145602  -8.3655118 -15.1970590   5.9570675  -7.6040882   2.7860390
#> [13]   3.5856114   2.2187777   7.2160881  -5.3198024 -10.3726846   1.5339374
#> [19]  -9.2802161  -1.2556306  -1.3156462  -1.1296278   9.6884234 -23.8812334
#> [25]  18.0640010   8.3102294   4.9593448   2.4565246 -29.6204175  10.0975236
#> [31]   0.8441359  13.4763332  -0.8075771  -1.4876285  16.3810124 -11.4813393
#> [37] -11.8146475  -9.9198354  -3.0268906  -1.2916250  -1.6027334   3.1142087
#> [43]  -4.6594153   6.0857978   3.4572189   9.5973511  -5.2705912 -10.5216322
#> [49] -10.2634743  11.1174387

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4876378