Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.069859049 -2.046838024 -1.059467025  1.782691090  0.806967884
#>  [6]  0.278057050 -0.431712038 -0.907397290  1.019162815 -0.384522024
#> [11] -1.448514492  0.101279551 -0.001262389 -0.835508530  0.800278601
#> [16] -0.704894047 -0.820006848 -0.378375308  1.019675498 -0.354378418
#> [21] -1.150605961  0.406455947  0.298923311  2.603381883  1.362150074
#> [26] -0.390404763  0.278788903  0.497295323  1.062956643 -0.609252343
#> [31] -0.730480217 -0.203503491  1.128814539  0.663362629  2.936442021
#> [36]  0.074024958  0.112567132 -0.185711833 -0.582252472  0.932810149
#> [41]  0.839029830  0.493576751 -0.901117695  1.954562663  1.248783736
#> [46]  0.211962229 -1.854485791  0.067667783  0.132600123  0.992583215

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.611168  9.726306 13.189695  5.343961 13.625041 10.277177 14.701503
#>  [8]  7.897334 10.763178 14.452800 10.984094 14.283887 14.921193 11.272717
#> [15] 12.703280  3.167393  9.622487 10.532749  6.525094 16.241368  9.100999
#> [22]  8.983110 15.919684 12.303475 12.705424  6.627320  7.309432 14.892503
#> [29] 13.553379 16.792350  9.866401  7.926545  9.148009 16.915604 17.122369
#> [36] 11.916076  8.632057 11.738950 13.272821  9.696121  9.349018 12.544844
#> [43] 14.867843 12.856114 15.612706  5.882942  8.931707 12.398111  4.435677
#> [50] 13.734305

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.1910796  -6.7811670  -2.3440089   9.5395724  -6.7649627  -3.5197334
#>  [7]   7.9117227  -5.3600457  -8.8540906  -8.5900591   6.6823355  -8.0717354
#> [13]  -6.1709076   9.9047497   1.4856937   8.1161241   3.1087650 -14.1682886
#> [19] -10.2886571   0.4607161 -13.0823872  14.8930481   5.2760766  14.6815443
#> [25]  -1.3175473   0.5138505 -13.0295648  -0.5697460  12.4701119  -0.8096506
#> [31]  -3.2573480   2.3019511  -6.1635034  11.0010904  -2.3177702  -3.7600580
#> [37]   5.8327972  15.1375835   8.3885971   2.2568658  15.3748340   6.1978627
#> [43]  -2.2089991   8.8169374 -16.4729436  11.0943061   2.2723257   8.3981069
#> [49]  -1.3776011  -6.8844060

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3140123