Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.66904626  0.43171599  1.59012271 -0.62279713 -0.58010400 -0.27507290
#>  [7]  0.21216481  1.78063983 -0.83394462 -2.01042446  0.49169007  0.63419609
#> [13]  2.23104932 -0.77237511 -0.82127675 -1.00520068 -2.83394360  2.43377118
#> [19]  0.02074714 -1.54423168 -1.19187278  0.63978452  2.86779291 -0.80559202
#> [25] -1.40633296  0.88650679 -0.52161301  2.90476565 -0.72415125  0.95133209
#> [31]  1.17307858  0.94682606  2.88369577  0.23311431 -0.21911579  0.85650159
#> [37]  1.32227480  0.83340035  1.67205219  0.59680547  0.11946754  0.56740856
#> [43]  1.29970061  0.74370884  0.05533554 -0.52917391  0.77907265  1.14418345
#> [49]  0.98231704  0.01445968

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.923576  8.473104 13.723859 10.030100 13.471881 11.849482  9.028343
#>  [8]  4.243694  9.117211 13.452767  4.847872  2.919163 10.124181 12.147244
#> [15] 12.330776  4.876064 10.177774 12.357531 15.914023 11.095685  9.189719
#> [22]  5.968437 13.175484 14.840143  7.571904  6.022082 17.339959 14.892788
#> [29] 12.784487 11.086434 11.561276 10.015376 11.176324 15.477575 13.831953
#> [36] 10.022930 10.887047 12.215535 14.491255  6.082513  9.444429  7.906683
#> [43] 11.817402 12.369854  6.702308 11.973320  8.522208 11.997916 11.466054
#> [50] 10.199927

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -14.2452289 -11.3813514  -2.5705459   0.6455679   4.4124379  -6.9341258
#>  [7]  -4.7272803 -16.3917039  10.3518305   6.0685584 -13.3555043 -10.3557348
#> [13]   1.6384548   8.0970815  -8.4233376 -22.6018768   9.9520469   4.7635993
#> [19]  -0.8305237  -0.2640642 -14.7051365  -7.6204610  -3.4950657  14.2837056
#> [25]   6.8097768 -16.8527120 -18.1945369  -6.3800221  13.0722973   1.6688580
#> [31]   4.1138888  -2.1009744  -4.5249000   1.4997163   2.4513908 -18.2983034
#> [37]  -6.1202943  -9.7768765  15.3907512  -3.4766138   3.6402592  -6.8485071
#> [43] -11.7379709  -1.1283826   5.4584030  -2.2697997   1.6304480 -15.3644176
#> [49]   4.1789763  -1.3804251

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2284957