Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.20441627 -0.92650545  1.44270127 -0.98259181 -0.24777529  0.89094709
#>  [7] -1.05235139 -1.71540569 -1.81993887 -1.96568842 -0.16026001  0.60018497
#> [13]  0.25352552  0.17036888 -0.31169492  2.60101148 -1.21516383  0.97303310
#> [19] -0.84184338  0.41732669  1.98169855 -0.26941399 -0.35585872 -1.60114532
#> [25]  0.63695680  1.66526018 -0.87961088  0.53897933  0.64298780  0.04463482
#> [31]  0.19955300 -0.56759948 -0.33314750 -2.63383340 -0.30537912  1.32498378
#> [37]  1.24086589 -0.90588508 -2.07431652  1.46374526  2.95288783  0.13098675
#> [43]  0.78454006 -0.88940519 -1.12161548  1.01111232  0.88955553  1.61523762
#> [49]  0.27740513 -0.16974633

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.145049 10.796800  5.228810  8.398468 10.856704  8.922393 10.280668
#>  [8]  2.518796  8.680525 10.881313 11.581219 11.951473  4.768571 11.472352
#> [15]  7.749372 11.213486  9.288124  3.951092  6.716886 13.060635  8.043381
#> [22]  6.341334 10.402174 11.025619 10.990584 16.013848  7.331806  2.585168
#> [29] 14.873426 11.196659 12.495583 14.108728  7.103633 10.278424 10.706676
#> [36] 16.651684 12.498194  6.622568  7.500783 11.839817  7.082957 14.744870
#> [43]  4.937534  7.879191 15.714018  8.512468  9.108037  9.838105  8.377341
#> [50]  5.086859

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.05068047  17.57213220 -11.27326061 -12.51654456  -6.96615487
#>  [6]   2.28681516   1.82892231  -0.89981842   0.03450674  -3.61964990
#> [11]   8.27639696   1.37545878   1.47966950  -2.01441202  -7.33879036
#> [16] -11.54438188   3.96620834   6.02161405   6.08596961   4.14317933
#> [21]   8.94786988  -2.61681189  -1.18524517  -3.14338836  -6.41649648
#> [26]  -1.08795303 -16.04077194  -4.69661488   0.10206548  11.59764980
#> [31] -15.11571391  13.55900232  -7.73518694   7.20754923  -9.24731063
#> [36]   9.32184381  -8.38273896  11.94489540   5.52433270   4.33415912
#> [41]   2.58032741  12.25684779  16.56450185   4.88509305  -4.90817462
#> [46]  -9.26099629   3.08774552   4.78851396  -8.30287214 -10.47983599

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5168791