Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.917630271  1.317300759 -1.082904988 -0.712800496  1.188741443
#>  [6]  1.011908499  0.636788471 -0.320223017 -0.016155750 -0.024580437
#> [11]  0.237733383  1.958260957 -2.698334910 -1.293079063  0.841629935
#> [16]  0.287688474  2.762103110 -1.221446308 -0.449784569  0.947887837
#> [21]  0.219293130  0.055126944 -2.222788314 -0.001330172 -0.434706564
#> [26]  0.364052267  2.100114878 -1.085319271 -0.370138824  0.299886816
#> [31] -2.229722082  1.350546805 -1.932659928  1.012871026 -0.321848586
#> [36]  0.246212459  2.430789261 -1.841870032 -0.200138656 -0.105495972
#> [41] -0.075885655  1.128395281  0.470650453 -0.715312910 -0.314868027
#> [46] -0.560185157  2.038319312  0.556300015 -1.454121104  2.169634753

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.740992 12.962570 10.665864 10.632124 12.028137 13.185220  7.400701
#>  [8] 14.964051  9.680576 10.800525  5.142858 11.636192 15.005713 11.526349
#> [15] 11.016460  5.823689  9.784724 11.384877  8.029643 11.627181  5.237840
#> [22] 13.976099 15.269423 10.047754  3.905702  7.280434 13.988369 11.788835
#> [29]  9.720047  8.834032  9.376562  9.987578  6.790288 13.567559  9.099465
#> [36] 13.766683 12.545676 12.463228  6.960956 12.577255 10.818059 10.785509
#> [43]  7.750625 12.105541  9.803137 13.786799 14.491626  5.799066  5.008006
#> [50]  2.254586

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -12.6850221   4.2727038   1.1600318  12.1769662 -17.7117940  -5.2300932
#>  [7]  -8.5492307   5.3695158 -13.7686085 -14.1345752  -7.9129516  -1.2576150
#> [13]  -1.8033554  16.3667327  -6.1892596   3.5293902 -11.6129171 -11.7838649
#> [19]   9.9475287   3.8953283 -10.6661306  13.5165200   6.3958452   7.7491293
#> [25]  -8.9771701  -7.8323158  -1.4536921   7.4613904  -3.9899184   0.1189970
#> [31]  -0.3560044  -3.5407043 -11.1809029   3.0224853   0.1098552 -11.1765028
#> [37]   4.8035918   1.9505049  -7.0484540  -3.1574981   6.5400783   1.7892776
#> [43]   1.1617107   6.6956327   0.8536940  -5.4920893   3.6193229  -0.2823312
#> [49]   2.5773850 -10.1490221

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2997885