Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.12358429 -1.28674957  0.29454634 -1.24201280  0.52225421 -1.61250602
#>  [7] -2.67962827  0.83356671 -1.67519784 -0.33458721 -0.14897324  3.34112503
#> [13] -1.43311629 -4.84521110 -0.07777650 -1.19515165 -1.52989316  0.56410038
#> [19] -0.76216154  0.80277459  0.35007102 -1.92591407  0.05284146 -0.36971403
#> [25]  0.21469195 -0.03777557 -0.55263170 -0.58771646 -0.79971304 -1.39491891
#> [31] -1.13602372  1.02636314 -0.27363796 -2.07008523 -1.33379015  0.15896648
#> [37] -1.60926259  1.45296082 -2.22147441  0.43642268  0.71254588 -2.48874349
#> [43] -0.91353586 -2.38997130 -0.22422585 -1.71232081 -0.72618551 -0.04488675
#> [49]  1.51405900 -0.33349737

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.522108  5.296982 12.205816  3.765376 11.169589 10.121112  6.705659
#>  [8] 14.166322  3.431926 11.807592  7.560166 10.611039 15.946205  4.388483
#> [15] 11.459386  8.665257  8.282775 11.018648  4.985867  9.210420 15.515339
#> [22]  9.484974 13.728635  8.559686 11.715056 11.090464  6.803517  9.498621
#> [29]  6.468031  8.387785 12.324295 13.043730 15.195916  4.740192  9.929922
#> [36] 11.030022 13.314796 12.867734  6.288869 12.318560  6.685646  8.472358
#> [43] 18.774478  1.086071 13.368207  6.787509  8.942721 10.165612  8.131682
#> [50] 15.212084

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   2.1042347  -9.8161083  -5.3926368   1.4251893 -15.7500857  16.3893960
#>  [7]   7.7910624   4.8091009   6.6560083 -11.0856915   1.9325457  -0.8296249
#> [13]   8.9137074   7.0362608  -4.2656818  -0.4565910  -9.5984041  12.6718629
#> [19] -10.9437508 -10.5182341  -3.5411979   1.6806470  -8.2487903 -13.0327962
#> [25] -13.0083514 -11.6937501  11.2698574  11.5438544  17.2160725 -12.7512554
#> [31] -14.2118427   2.4341655  -7.6037577 -13.4434426  -1.2385849 -18.7200520
#> [37]  -7.2243199  -1.0187845   0.4053473  -0.4123329   4.8377369   4.3541283
#> [43]  15.7064629   6.3636282 -14.6849867 -13.8394381  -4.4840343  -1.5258601
#> [49]   6.0195156  14.7123261

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5300043