Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.4125587  1.5563251 -1.1173475  1.9784784 -0.2937836 -0.6867505
#>  [7]  2.6489236 -1.5162686  1.9966195 -0.9418162  0.4277473  1.0273164
#> [13] -0.9938952  2.0386574 -0.7576829 -0.6419742  2.1324331  0.9146638
#> [19]  2.3057984 -0.9875230 -2.8358223  2.3198233 -0.5508128  2.7849398
#> [25] -1.9281687 -0.5427822  1.1626360  0.6247490 -0.8010020 -2.0516206
#> [31] -1.2348531  1.8752674 -1.6883267  1.5582959 -1.0183028 -0.9050502
#> [37]  1.0407221 -1.4537547  2.1350485 -0.8238492 -0.2220821  1.8045901
#> [43] -1.3569786  2.0860423 -0.8567987 -2.4282258  2.2425998 -1.6938310
#> [49]  0.9021480 -1.8760640

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.412417  7.173611 13.668289 10.398103 15.153311 15.164815 12.617271
#>  [8]  7.606468 13.037319 12.377322  7.763507  6.505673  9.621824 16.498241
#> [15]  6.051076 15.875513  9.341293 10.379499  9.370351 12.347795 14.155385
#> [22]  5.048454  9.681317 11.534437  6.977803 11.556417  5.970598  8.609072
#> [29] 19.237725 12.346156  9.227498  9.345810  5.580212  9.811285  9.428145
#> [36] 12.919403  6.585981 12.281367 16.624768 12.481296 11.929330 13.341227
#> [43]  5.738539 13.420027  6.533193  9.654094 10.585984 13.113731 14.098878
#> [50] 10.422351

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.8548725  -0.4281707  -2.9286639   4.9280364   4.2043881 -11.1406560
#>  [7]  -5.9059784  13.5072345  -1.7474543  -1.7324555  -8.7248763   7.7489781
#> [13]  17.7044600   4.2362752   0.3871047  -6.7913601  14.7768586   5.1613659
#> [19]   5.1223245   7.1422456  -0.9780402   9.1141187  12.6866629  18.4564484
#> [25]  -1.6599927   0.1978466   9.5583874  10.2589207   9.4077889  -3.7070923
#> [31] -10.1287536  -2.1516586   2.5481500   1.4414173  -4.5188538 -11.0687136
#> [37]  -1.8641182  -9.7952094  14.0850975  -3.8496028  10.8525444  -0.7015201
#> [43]  -2.1916257   1.1974682  15.3581619   1.4561960  -3.5475106  30.9357712
#> [49]   5.1937213  16.7505534

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1844105