Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.18661614  0.61988910 -0.60391433 -0.15454791  1.09224010  2.70510641
#>  [7]  0.26613284  0.11918102 -0.95432642 -1.68526364  1.10913264 -1.89332583
#> [13]  0.41020191 -0.01152392 -1.66187190  1.85438894 -1.00770078 -0.11596872
#> [19]  0.96171810  0.67244966  2.45093480 -1.15926306  1.14741876  0.99928199
#> [25] -1.65655436  1.34391537 -1.10912897 -1.06204237 -0.74443005 -1.20486592
#> [31]  0.52377911 -0.83335671 -1.11244684  1.39944549 -2.42753390  1.43650894
#> [37] -0.97064967  0.14386098  2.51727521 -1.17461254  2.35780327 -0.52987321
#> [43]  0.44340037  0.51517410 -1.02145885  2.00549347 -0.57852493 -0.33495835
#> [49]  1.71337288 -2.82053051

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.062801 13.508886  9.642477  8.945700 12.044204 15.157880 10.059636
#>  [8] 11.422077 10.971852  9.088798 15.951633 11.155168 10.166916 10.215177
#> [15] 13.313313 11.972835  7.409938  7.808851  6.617469  9.406943 10.474481
#> [22]  8.414405  9.061066  5.910993 12.502057 10.793829  9.401747 10.995539
#> [29]  4.297041 10.963903 13.164720  5.867511 15.062192 13.992107 18.741692
#> [36] 10.186660  8.758183  7.374812  3.685382 12.751052  5.803075 10.375244
#> [43] 16.833084 11.230484 12.211386  6.924590  9.433149 10.955471  9.000190
#> [50] 12.665988

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  14.6792679  -8.5260032   4.1783615   6.2553297 -12.2128702   8.0166024
#>  [7]  -4.4191305 -10.1479728  -3.7371215  -7.3611076   5.3029153   1.1849499
#> [13]   9.9254271  12.3299652  -3.2549325   6.2865460   0.7847885   5.5203980
#> [19]  18.6586045 -23.7359642  -9.2461859  -9.3554504  -4.9178692   2.7609685
#> [25] -20.7303353   4.0428864   0.2569860   7.2112916  -7.4185884 -13.1262177
#> [31]   4.5337325   4.2328571  18.5553524  -5.2055151  -0.8931021  -4.4244599
#> [37]   6.4757299   4.5594121  -7.7596440 -13.8017251 -14.6022231   6.8671866
#> [43]   6.7833816   8.1863837   4.9160400  -5.2347302  -7.0350039   7.8381002
#> [49]   2.7025879   5.7749242

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3872067