Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.03229220  1.54772940  0.99844588  0.06214200  1.02244505 -1.89538453
#>  [7] -0.39794605  1.26186653  0.46926440  1.92670694 -4.45898523  3.15605244
#> [13]  1.62326308  1.28585983  3.16312612  0.07887672  2.45331579  1.80010806
#> [19] -0.43763080  2.29824986 -1.94186896  0.44880691 -0.39654368 -0.72996620
#> [25]  0.94047006 -1.31144531  2.69866405 -0.04146414 -1.43675248  1.61912583
#> [31] -1.27953705  1.39633920  1.43588380 -0.76689143  1.55625021 -2.08178289
#> [37]  1.64194277  0.86909339 -0.35366374  3.65055877 -2.48649480  1.19421178
#> [43]  1.30257107  0.11643192  1.91963239 -2.57881487  1.59898931  2.23994319
#> [49] -0.52808343  2.51302568

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.429947 10.983152 12.150541 13.071875  3.973186 10.330277 14.044206
#>  [8]  7.407812 17.154969  8.903469 13.772140 16.110501  9.612448 17.472701
#> [15] 10.632970  8.583670 16.287172 10.206273  7.456287  8.231871  9.481258
#> [22]  5.472224 10.966000 11.762427  9.191457 23.056751 11.467042 12.533519
#> [29] 15.275292 12.138645 16.570097  7.109293 10.923185  9.211481  7.037936
#> [36] 14.445068 15.980899  9.853782 15.050469  9.385319  8.581920 10.013441
#> [43]  7.822677 10.029492 12.872309 18.617735 10.243416  7.769379 13.570111
#> [50] 12.002687

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  10.596357557   6.080440165   0.439242098  -3.336200616   2.947762343
#>  [6]  -1.858613897   1.137238404   1.368839920   7.781063516   4.102135254
#> [11]  -2.307413678   4.208706990  -2.923102557  -2.607035318   9.700141324
#> [16]  -7.586614833   6.477576853 -20.680211707  -7.216167987   8.314329894
#> [21]  -7.097664647   3.843181647  -8.367022824   3.057559130   0.608338515
#> [26]  -5.451492078 -10.468603112   0.649287582  12.311317067  -1.544187052
#> [31]   8.176245466  -2.051068472   1.895102977 -20.432041840   2.902288445
#> [36]   7.482499708   1.284063158   8.039166079 -10.216654409  -6.340953892
#> [41]  13.401653767  -0.070505393   7.080531965  11.533705505  15.960460782
#> [46]  -0.366087460  -1.193098386   1.279240402   0.005661666  -5.189961725

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4827751