Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -3.64452297 -1.30113979 -0.60193394  0.29364705  2.94577767 -0.19562187
#>  [7] -0.36125948  1.05335467  0.71588436  3.23586396 -1.31898176  0.48986801
#> [13]  0.44705574 -0.56639045  1.40171388 -1.66097996 -0.39862122  0.45337538
#> [19]  0.65769410 -0.35971680 -1.56738460 -0.71554909  0.95754734 -0.28056753
#> [25]  1.32945700 -2.18661071  0.17443456  1.53794860  0.34414301  1.02725930
#> [31] -0.05897976 -1.84486984 -0.75285067 -0.17081015  1.60681467 -2.10717876
#> [37] -0.75993699 -1.93960625  1.02068110  0.57307712 -0.97070269  1.20899506
#> [43]  1.63400032 -0.96445620  1.62347320 -1.93051888  1.01308944  0.09515380
#> [49] -1.29744786  2.93004126

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.899147  7.005995 10.216899 11.645000 12.023726  6.304408 14.832291
#>  [8] 14.420109 17.112954  7.657843  6.222866 12.436362 12.712959 12.483661
#> [15]  7.181936  5.775421  8.949808 10.818944 11.732642  7.524396 10.867482
#> [22] 13.027597  8.100555 15.991140 10.132715 11.570952  8.416769 16.396579
#> [29] 10.195215  6.944989  3.685470  9.832687 14.191794  8.608944  7.909302
#> [36]  8.213632 10.186483 21.038811 15.819102  7.953919  8.630055 11.150240
#> [43] 14.172310 11.797000  6.595779  5.720665 13.476542 14.125934  7.418805
#> [50]  8.084870

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.45522473   0.58814904   0.56313606 -16.78027300  -0.03431987
#>  [6]  -6.16546255   3.72347599   1.23634916  10.27477707   4.36841255
#> [11]   2.95016595   4.27845333  13.12602530  -6.11291194  -2.15965608
#> [16]  -2.69735653  -8.07638206  10.32769903 -19.82721276   4.62042840
#> [21]  -0.01842188   5.29124294  -1.52236171 -14.40128623  -7.71503166
#> [26]   6.15099523  -9.93319530  -0.60035516 -18.90755122 -12.59288971
#> [31]  -0.14524901  -7.23869116  16.30271713  -4.85308716   0.09978224
#> [36]  11.31429726   7.74179323  12.40124953  -9.93003844 -17.67016617
#> [41]  -5.65779458   3.75743869   4.79608618 -10.62856037 -18.93331766
#> [46]  -5.88017830  -2.25434922 -10.39401043   0.47246359   3.68377934

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2086765