Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.298117992  0.968047623 -0.431367088  1.796260360  1.974310724
#>  [6]  1.681408516  0.591894464  0.427693808  1.703734418  0.574570506
#> [11]  1.500484584  1.421592303 -0.114795889 -0.842979685  0.003525049
#> [16]  2.463702325  1.004018408 -0.890770710 -1.350147146  0.739837638
#> [21]  2.364677662 -0.235163232 -0.054712835 -0.325376149  0.098121670
#> [26]  2.200768767  0.508670838 -1.251391812 -0.363236661 -0.019308994
#> [31] -0.833368090 -0.342183371 -0.412142378  1.724874827  1.484818668
#> [36] -0.661020035  0.189133881  0.403204127  1.215857351  2.305371508
#> [41]  0.997466386  2.104358189 -0.019717442 -1.363352175  0.907335395
#> [46] -0.302030528  2.129080838 -0.232477689  0.451800478  0.009507712

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.971098 12.105062  9.101848  7.418003 17.311551  1.865133 14.634711
#>  [8] 14.083396  7.156201 13.419910  6.881910 12.152215 14.766766  8.201292
#> [15] 13.934912  9.861160 19.183450 11.407697 10.136062 12.036523  8.757691
#> [22]  9.173610 11.018389 10.698093 13.675341 13.469777  5.473982  8.442721
#> [29] 12.597987 11.261125  3.780472 12.212512 12.309163  7.498847 13.039565
#> [36] 11.881536  6.294559  5.319090  8.103620 13.814603  6.466536 12.153751
#> [43] 11.348406  5.534450 18.411037  9.921056  9.467485  8.991754  5.477231
#> [50] 17.248350

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  12.9494934  -3.3784952   6.1493874  -1.5763445  -5.0317503   7.5856983
#>  [7]  18.0508079  -9.1990354  -2.2136247  -8.1546491   6.1445297  -5.3381792
#> [13]  -4.9562098  -4.4202926 -13.9787894  -0.2282487   7.1163944   7.1216120
#> [19] -10.9577340  -2.0229002  -3.6915221  -0.1447102 -12.9089777  -6.2823962
#> [25]  -8.2846763   2.6915667  10.5378268  11.0208097  -8.1242323  -2.3678378
#> [31]  -6.5289288   3.6282368  -5.3695216  -6.3916868  12.2102469 -12.8951657
#> [37] -13.9143564  -9.7642709  -1.0092305 -16.1166018 -10.2019966   8.2775048
#> [43]  -2.8565018 -20.2416495 -11.8659557   8.5344176  17.6394530   9.7193721
#> [49]  -2.7007375  -6.7867124

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4622019