Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.779098063 -1.117947228 -1.245710929  1.452496365 -0.676165694
#>  [6] -1.302562249 -1.089718172 -0.456859345  3.408436507 -1.261807092
#> [11] -0.012902204 -0.009888903  0.290705980 -0.015997153  0.016227993
#> [16]  0.388838110  0.019016725 -0.631192381  1.086036265  1.064520303
#> [21] -0.229965551 -1.637811084 -1.310149625 -0.211567812 -1.200779645
#> [26]  0.093740765  0.649304812 -1.840884894  0.617523611 -0.882011114
#> [31]  2.029597532  0.164525024 -1.149241962 -0.461657431 -0.497902701
#> [36] -1.979147633 -2.196579384 -1.502870842  0.264169385  0.855486766
#> [41] -0.052522248  0.099708616 -2.304614165  0.496567267 -0.421978781
#> [46] -1.761459485 -0.328358153 -1.205002023 -0.285179727 -1.251948406

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.6890250  8.5727109  6.7224642 14.5227658 10.0714198  3.6087265
#>  [7] 12.0548035 10.2616224  8.1670340 11.1275400 11.4349364 11.0579956
#> [13]  6.1568013 12.1153208 13.4200389 13.5831563 14.3906994 11.9824860
#> [19]  9.9124860  7.5839104  0.1798755  8.5857723 11.4032707 10.6950939
#> [25]  6.5996343  7.9927480 13.2755610  8.2090285 13.9288033 10.0686525
#> [31]  8.5889614  8.6687312  6.2046494 12.6800119  3.6428653 10.3566690
#> [37] 10.5945161  7.8920628 11.2260391 11.4299831 10.8130662 13.1607481
#> [43]  8.7133873  9.7651379 10.6699662  6.7771557 11.1890866 13.8639087
#> [49] 13.6739204 12.4254306

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   6.0662698  -4.3620290   0.8963491  -2.9092800 -14.1430729   0.9282611
#>  [7]   3.4377070  -2.9698251   1.9272395  -1.0771756  -9.9613995 -19.3162294
#> [13]   0.8489736   8.0499127  -9.3292306   1.2318097   5.5407783  -4.0304893
#> [19]  -7.2554847  -6.9809542   6.1035782  19.3761321  -0.3926685 -18.6546656
#> [25]  15.2878324   1.1030781   6.1882528 -10.1597462   1.3641653  -1.9289527
#> [31]  -3.8272449   7.0413737   2.7941861 -15.3961866  -2.0463526  -5.5281352
#> [37]  -3.0411487  -6.5361107   5.5199249  -4.7021585   4.2390944  -7.1547882
#> [43] -12.8743725  -7.1258859  -5.0103508   1.2363784  -6.9777966   0.2039258
#> [49]  -3.2136964  -4.3980041

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5713816