Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.01379046  3.64413477 -0.43130815  2.81424730  1.44989125 -0.94757479
#>  [7]  0.78372344  0.82709633 -1.72189939  2.43654851  0.29009953 -0.38504024
#> [13] -0.16404358 -0.59796515  1.82491650 -1.88499521  1.54924204  0.23002065
#> [19]  0.18418149  0.71133374 -0.84076129  1.75587357  0.71051695 -0.69619438
#> [25] -0.21142890 -1.49883570 -0.41338947 -0.65692229 -2.05473107  1.39349485
#> [31] -0.67574125 -0.59663155 -0.39989542 -0.50352506  2.58915509  0.48604409
#> [37] -0.66254908  0.33102615  0.84652546  0.18500684 -2.39249945 -0.86769137
#> [43]  1.08938692 -1.09344356  2.09128830  0.15780138  2.19547284 -0.14677746
#> [49] -1.60432602  1.70078993

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.9291224 11.7040783  8.5099615  5.2260679  5.5781342  0.5959818
#>  [7]  8.7160288  5.0955975  8.6271114  5.8080860 12.2546852 12.6172886
#> [13]  8.7365761  8.2880184 11.3587334  7.5914053  2.4989812 12.7388135
#> [19] 11.1782640 11.0539863  1.8866925  9.6820072 11.5146489  8.4116023
#> [25]  7.7051691 12.5781427 10.5972103  8.9951485  9.5488206  7.6554272
#> [31]  4.3880241 14.9757587  9.9304193  6.5235044 11.6603970  8.0854285
#> [37] 11.7043787 12.7241433 13.9872333  9.1560403  9.3996625  9.2203382
#> [43] 10.0099793 13.3968545 14.1747200  8.7676095  9.2121026 10.0519928
#> [49] 11.6717690 11.8969960

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  11.5438452  -8.0531976  -1.9125965  10.8193118   4.5646900  -2.3787196
#>  [7]  16.3905405  -5.1109038   5.3990651   7.2662109 -26.8991562 -12.0605802
#> [13] -11.9900585  15.3745322   5.1280372   7.4572806   3.1320175  -3.5429150
#> [19]   7.2659068  -7.3341202  -2.4220145  -3.7972321   3.1305341   5.8548622
#> [25]   8.7854728  -3.2509371  -1.5697673  16.8002587   5.0714960  -1.0102592
#> [31]  -6.3221218   2.8638762   1.7814642  10.9863171   7.7064500 -21.7532404
#> [37]   0.6678019   5.1340536   6.1694933  11.4059188 -14.9967284  -2.3739766
#> [43]   4.8704706   7.6252123   2.3169798  -8.2213568   0.8547046  20.3741522
#> [49]  -3.7827330  -2.7681299

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3719819