Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.42842610 -1.25934560 -1.61486383  0.09120079  0.80097071 -2.11289616
#>  [7] -1.13307732  0.27339507  1.91347661  0.31167985 -1.89561981 -2.23023707
#> [13] -0.79644921  2.99673041  0.63904677 -1.40600164 -1.18156615  1.27570303
#> [19]  0.68431353  0.65465341 -0.72976190  0.04854416  0.41907197 -0.35386063
#> [25]  0.24783702 -1.60600000 -0.48743453 -0.08160813  0.51887932  0.05404038
#> [31]  0.38841941 -0.87825653 -1.16524648  1.22274226  0.22586708 -1.24874182
#> [37] -1.34537396 -1.62348575  0.45837360  1.25553866 -2.23602285 -1.35881566
#> [43] -0.51392508  0.34358467 -2.10325312 -2.27506950 -3.58017843 -0.78082595
#> [49] -0.28783008 -0.40741795

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  2.883061  9.071735  6.669357 11.423304  9.793531  3.296206 10.800951
#>  [8] 13.351672 12.476388  9.499297  4.967806  7.434736  7.035145 17.076540
#> [15] 13.738286  8.387838  5.289694  4.842123 16.412448  9.537006  6.477964
#> [22] 14.154601  4.069860 16.589449 12.306704 11.209579  9.825196  7.205745
#> [29]  7.858871 10.784652  6.881457 11.217940  8.042393  8.700643  8.938230
#> [36]  8.309094 14.118413  8.954655  8.726321  8.838824  5.756673  1.490006
#> [43] 12.155724  8.030548 10.529619  7.077741  5.769287 10.545093  9.346813
#> [50] 13.088789

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  10.51992942   6.13313668  13.27409548   8.94609363  16.40730049
#>  [6]  -3.81296264  -1.83653261  -5.72891845 -13.26909634 -13.51233987
#> [11]   9.04231686 -15.20260926   4.45821095  -7.00469806   7.51981765
#> [16]   4.82430033  -2.53211781   2.85560113   1.77324230   0.01728250
#> [21]   2.50162579   3.83180341  11.20061242  13.85059090  -1.39277830
#> [26]   8.02910394 -10.75731020  -8.11427690  16.27519467  -0.32185859
#> [31]   4.90167268 -11.74197413   5.65151649 -12.82304004 -10.42933669
#> [36]  -6.49897299  -7.92061989   0.04605174  -6.13444884   1.24907897
#> [41]  12.28375720 -15.12669944   7.98091714  -1.59756549  -3.95502455
#> [46]  -3.13714109  -7.30649441  15.16086538   2.80143828  17.70466656

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1855986