Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.06961784  2.19104542 -0.09490755 -1.87386321 -0.69561169  0.38749395
#>  [7]  0.16030186 -1.19145990 -2.12317344 -1.41682145  0.14764791  2.23674161
#> [13]  0.23091815 -1.98466069 -0.46535814  1.30177463 -0.06965753 -0.14161695
#> [19] -2.22320943 -1.95754152 -0.26899217  4.26366784  1.39924235 -2.48429176
#> [25] -1.48913918 -0.06022695  3.45659315  0.63480451 -1.65041172 -1.11546628
#> [31] -1.70788338  0.14913631  1.94836057 -2.71514695 -1.47321818  0.96809025
#> [37]  2.16358402  0.23359741 -2.63360528 -0.31273900 -0.47665519  1.40562108
#> [43]  1.17625032 -0.03081729  1.12313612  1.31735046  1.98238413  1.07456178
#> [49] -1.54162641  0.18765945

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.991141  6.314987 13.092504  9.189462 12.345644  9.474798  8.918924
#>  [8]  8.722487 10.791446  4.606065  8.319157 13.736122  8.047235 13.882648
#> [15]  7.438923  8.139428  9.293329 19.106115 11.313706  5.302708  8.242586
#> [22]  5.071668  9.307772 10.571145 12.459821  1.629139 10.166332 12.818392
#> [29] 15.751306  5.288146  7.229224 11.692390 12.464855  8.852188  3.577597
#> [36]  4.122485 11.325744 13.734762 11.367342  8.312311  7.911107 12.386786
#> [43] 14.513128 13.707125  1.272295  4.499885 14.860452 13.258205 16.929074
#> [50]  7.634454

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   7.45732044 -10.94257037 -13.42966612   4.86856817  19.09654214
#>  [6]   9.47955784  -7.80403547   1.48021642  -6.08638774  -1.27198286
#> [11]  -5.41908080   1.18241305   0.81534293  -4.04459723 -12.57897844
#> [16]   3.05435603   4.35546589  12.17321000   0.06950821  -3.14777554
#> [21]  17.66247308  -5.38604052  11.32433414  -0.13700602   1.05359147
#> [26]  -4.70157161  10.05799735   1.04806605  -2.70157318 -14.62718923
#> [31]   5.91901688  -3.27718198  -2.63508683   9.49572598 -17.67348162
#> [36]   9.92000070   8.08854061  -3.48779539  -2.54601017  -1.55555685
#> [41]   6.50509460  -8.78924840  -4.38365464  -9.93285625   1.93337095
#> [46]   3.47956943 -12.10603227  -2.71863105  -5.78907151  -8.92169023

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5285315