Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.9485537  0.6495315  0.9887677 -1.5603308  1.1954578 -1.4348638
#>  [7]  1.4038399  0.9211811 -0.2907625 -0.1225077 -0.4844660  2.3285648
#> [13]  0.7455190  0.6264620  1.1344838 -2.1525810  1.9278133  1.1045064
#> [19] -1.7145357 -0.3165376 -1.8942894  0.3351735  0.9843629 -0.5590339
#> [25]  1.3442073 -2.4384945  1.8076544  0.6574772 -1.0572342  0.5204092
#> [31]  0.3857391  1.1890086  0.8786908  0.7467699  1.7173760 -1.2290062
#> [37]  1.2405772  2.7536730 -0.9741976  2.2357941 -1.2833475  2.6155242
#> [43]  1.4458084  0.3476442 -0.2351502 -2.1723018  1.7373165  2.3772094
#> [49] -2.1378833  0.7123717

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 16.34876279 10.27451702 15.58307378 13.15701231 12.18693283  8.90650754
#>  [7] 10.24892713 11.23898626 15.25262499  9.59155491  8.21988182 12.06107602
#> [13]  8.92452854 17.86729950 12.15836279 10.73774575 17.59399284  9.93923105
#> [19] 16.58460918 12.21433376  9.61793118  8.44825998 13.92751865 13.48768108
#> [25] 12.87160497  2.91750684 10.34091660  8.29014325  6.59248590 14.41710489
#> [31] 10.18282152 10.68436194 11.13845003 14.92604441 12.48940996  8.03949418
#> [37] 13.21788008  8.79831916  9.32879635  8.75710056 -0.08147852 10.70551952
#> [43] 10.59880174 10.89117916  9.51373362 11.66878496  9.82012672  8.72103108
#> [49] 11.48381620  8.78290687

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   2.8681102   2.9675389   4.1621796  18.4351466  -4.9228144  15.6870416
#>  [7]  -5.5596148  14.8967527  10.5187664   6.2818510 -10.5531658  -1.9337492
#> [13]   8.7434009   3.5418218  -7.7916368  -2.4364896   0.5055800  -5.3320936
#> [19]  18.3996517  13.9744484   1.3484774 -14.8530176  -1.6778706   7.6182785
#> [25]  -7.9120303  -7.4563933  -5.3606203   7.8879835  -7.3719322  -4.9229198
#> [31] -12.2784529  -1.0083182   2.3993922  -9.8621151  -5.6718220 -17.1076964
#> [37]  -2.3776305  -4.4247961 -15.9023926   4.6291463  12.4000516 -15.5699317
#> [43]  -3.2307416   7.8154756  -6.0666650   6.5808037 -12.9548492   8.3821985
#> [49]   6.8923649   0.4343888

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.240946