Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.984038652 -1.060236023 -1.861760121 -0.214372751 -3.272116925
#>  [6] -1.659401184 -1.480253343 -1.865183334 -0.711189109 -0.991600606
#> [11] -0.284364284 -2.775810092 -1.019524589 -2.183452284 -0.971830044
#> [16] -1.968644190  0.201866678  0.038368892 -1.164225118 -1.659488765
#> [21] -1.210538365 -0.008638792  1.259413795 -1.449342159 -1.818917376
#> [26] -2.042017088 -1.947745317 -0.554179537  1.452624986 -3.480307807
#> [31]  0.077927510 -1.239787731  0.778571612 -1.127384373 -1.148684458
#> [36] -1.473940653 -0.241616936 -0.537739178  0.152017866 -2.183608213
#> [41] -1.128499285  1.452647174  1.081429842  1.219590668 -3.308212878
#> [46] -2.331582161 -0.156041986 -1.798078449  0.049600088 -2.075083107

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 16.380307 17.282275  4.347993  8.792185  9.708346  7.358569 10.326933
#>  [8] 10.991831 13.549179 15.835259 13.427166  7.378043  8.136787 10.995088
#> [15] 11.380043 12.563253 11.691632  3.721035  4.151423  7.802710 11.770505
#> [22]  9.573095  8.089518  3.541196 10.981589 10.538670  8.770682  7.203000
#> [29] 13.125462 11.257376  9.218126  9.987996  8.531900 11.438185 10.128455
#> [36]  6.158254 15.517041  9.475750  8.988223 10.189294 15.638452  9.921556
#> [43] 11.140791 12.053882 11.489144 10.397123 13.776043  4.794467 13.043661
#> [50] 12.379891

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -11.7810844  17.0405841  10.1080637  -2.3063823  14.2701128 -12.1643804
#>  [7]   0.6863270  11.2976249  -3.9862519  -1.9762508  -7.4784621  -7.1048236
#> [13]  -2.7083119   6.5041980  -5.1327447  -0.3822550  -1.5759522 -12.9724389
#> [19]   1.5638621  -0.6604623  -4.4587222   6.5974121  -8.5832564 -12.4484636
#> [25]  -3.9764969 -14.4392302   7.3708292   2.8774281   5.1325280 -12.5606756
#> [31]  -3.6739846 -24.1599681   2.6772495   5.9643822   3.9054386  -5.8384537
#> [37]  12.3146666  12.2303055  -6.0057588   4.8740316  -6.0282522  -0.7658569
#> [43]  -1.1563597  -9.5951240  -2.2182215  -9.9492247  -1.3720496  -1.7240317
#> [49]   0.3057999  -9.6371335

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2753809