Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.8306255 -2.0252446 -2.5914021 -0.4425404  0.1313861  2.3832405
#>  [7] -2.9489458 -2.3113081 -1.0373744  0.6334864  1.9039994 -0.4638652
#> [13] -1.7631819  0.1949319 -2.0275148  2.7733523 -2.3502264 -2.6489495
#> [19]  0.4603167  0.3484759  0.4043448 -2.4192756 -2.7690896  0.2794642
#> [25] -0.4784042  0.4219833 -0.2761770 -1.0109583 -0.9124882 -1.2991592
#> [31] -1.4012264 -1.5698459 -2.2684560 -0.9763650 -0.9320455  1.6821957
#> [37] -1.4331569 -3.8329258 -0.9780054 -0.5287169  0.7304894 -1.5445258
#> [43] -0.9640890 -1.5441756 -0.5270407  0.1718776 -2.0971106 -1.7867282
#> [49] -0.3078965 -3.5921611

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.161307 16.942360  8.782129 10.951442 10.373516  9.432329 11.370436
#>  [8]  6.913662  7.064998 11.488422 10.621194  9.265198  6.608384  9.554209
#> [15]  7.404344  6.882269  8.465915  9.874690  4.285275 11.569429  6.826256
#> [22] 14.193600 13.344135  2.172989 11.440580  9.046065 14.112877  9.005569
#> [29]  5.547667 11.975805  8.406428  7.453851  6.433346  9.983157 10.681606
#> [36] 11.953003 13.561770  4.097341  8.898877  8.746897 12.060370 13.698889
#> [43]  7.609935  3.924369 15.170309 10.666625 11.718333  3.100440  6.122505
#> [50] 13.684666

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.7827083   0.1025397  13.2562856  -9.3731440  11.3736613  -8.6096796
#>  [7]   5.8827105 -14.8080818   0.5914431   6.8451676  -5.3820823   8.7282316
#> [13]   4.4962304  -5.8338280  -8.9895646   6.4007202   6.5308408  19.3867041
#> [19]   0.3239377  -4.3349002  -0.4535203   9.1888971  -6.5747153  -9.0917863
#> [25]   1.4182878 -10.1656004  -4.3274929  10.5735385 -10.3114759   4.7535441
#> [31]  10.3361436  10.4528490   1.1839327   3.6355863   0.8395665  -2.9921939
#> [37]   0.5159908   2.9450026  -9.3528200   4.1106982 -17.0012760  -0.8902637
#> [43]   6.2955054   0.8865996   8.0909467   2.2288026  -1.8522958  -1.0643648
#> [49]  12.2734426   0.7221428

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6433773