Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.07880586 -0.61714188 -1.03795174 -0.30390717 -0.29516529  0.70613138
#>  [7]  1.98337623 -2.86708634  1.32111448  0.29586369  0.05983759  1.79741708
#> [13] -0.64718242 -0.78220749 -0.47995344  2.96985684  0.70262106 -1.41570194
#> [19] -1.90706965  0.63844500  0.46123234 -1.57162231 -1.03155661  0.28625952
#> [25] -0.39885801  0.34625512  0.13860091 -0.53360042  0.16735263 -0.63049371
#> [31]  1.27790372  0.83804429 -1.71093799  0.34558154 -0.14080214  0.33518492
#> [37]  0.34751662  0.21376848 -0.64623468  0.31268401 -0.15145118  0.68142871
#> [43] -2.84889864  1.38376692  0.95344669  1.91232602  2.40985210 -1.23228093
#> [49]  0.31450303  0.35435275

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.564385  5.840340  8.042090  5.210500  5.328438  6.636177  6.747758
#>  [8]  5.375116  5.000856  3.733014 17.002688 13.415420  8.542808 10.717058
#> [15]  9.004909  9.723708  8.410529  3.234604  5.216056 11.090057  8.338452
#> [22]  8.935139  9.352258  3.317355 11.318481  9.003847  8.069360  8.981920
#> [29]  6.828894 12.271347  6.292695  9.675080  7.997432  9.716518  6.069107
#> [36]  2.914984  8.957888  4.623744  7.323581  4.770236 12.036540 10.897167
#> [43]  5.136771  8.957036  6.177278  8.481718 10.819223  7.349950  9.461701
#> [50]  5.581783

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -8.96371147 -24.44611117  14.39360335  -3.29800338  -1.71438262
#>  [6]  -0.80132315  -2.02514059  -2.28324064 -14.79582757 -10.34017569
#> [11]   1.17628230  -2.39835651  22.41169614  -6.55155531   6.23036274
#> [16]  15.64758286  14.38067718  -2.62622155   1.85251650  -5.54021018
#> [21]   4.26844732 -10.48900316  -0.26002284   8.10967368  -3.81791212
#> [26]  -2.68308594  -1.47155420   6.07786708   2.30750623 -10.94519088
#> [31]   0.48324183   3.05515436  -3.19881852  -2.30442338 -16.11451719
#> [36]   6.91252366   1.43199243  -0.05580531  11.27013013  -3.84992157
#> [41]   6.46362387  -0.75417721  -5.06100674  -5.30007976  -2.60569158
#> [46]  -2.19498370   0.54117851   4.05308982   8.89147423 -11.13038992

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2944928