Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.037867723  0.001340147  0.841992045  0.391680917 -0.919990030
#>  [6]  2.055838727 -1.068176352  0.024696046  1.707566464 -1.752084116
#> [11] -0.369259703  0.640652020  0.090527005  3.028010683 -0.983683686
#> [16]  0.389449690 -1.266085811  2.114183603  1.760469711 -0.506280669
#> [21]  1.898415975  0.612040609  0.960815898  0.397350496 -1.061395285
#> [26]  2.291152771  0.183096057  0.809417376  1.897858796 -0.570344962
#> [31]  0.519212493  0.271056987  2.066627913  1.993418379 -0.543031629
#> [36]  0.479151232 -0.181209630 -0.465482423  2.600077673  1.317753951
#> [41]  0.061587356 -1.318063974 -0.856600577  1.790340795  0.850737643
#> [46]  1.451504811  1.758243270  0.286370992  0.814482410  0.001656947

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  5.647542 10.677274 17.819617  9.775882 17.805886  3.877542 10.138748
#>  [8]  7.953862 11.037320 11.835995  3.749731 12.147917  8.337644 11.297478
#> [15] 14.130225  6.100554 17.106214 14.204150 13.688166 12.957007  7.641600
#> [22] 12.765327 10.267979 14.168274 14.073416  8.174545 10.846053  7.012012
#> [29] 14.772252 15.024696  8.349583 11.402466 14.640391  8.113762 10.662143
#> [36]  5.391867 15.503991 14.693857 16.211984 12.159427  4.811443  5.979192
#> [43]  8.120177  8.007573 17.559612 10.540393  7.678497 10.293845 10.548857
#> [50] 12.336130

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.79582981   4.06007953 -10.73971356  12.91498351  -8.16672072
#>  [6]  14.37262570   4.77337371  -6.28578444   2.87244244  -2.62911251
#> [11]   8.04170246  -1.90706417   6.07565131  -0.09086025 -20.92285508
#> [16]  -0.49813656   3.66790352   1.59664018  -3.14616065  -3.61061493
#> [21]   5.67906684 -24.12151167   3.27550919   3.87012789  11.13522605
#> [26]   5.25808725  -8.70640278  -0.06878070  15.89294317   3.61803224
#> [31]  -1.53497846  -6.17660427  -4.31445788 -10.29436834  -3.99393206
#> [36]   5.42729485  -9.40052868  -8.04483381  20.16637642  -0.47209353
#> [41]  -5.86781647 -12.59386579  -1.60916127  17.60276705  -2.70523533
#> [46]   1.64531732  -2.74579022  -1.89313365   4.37442773   2.80413559

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.539208