Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.17154826  1.08260687 -0.79229850 -2.87652777  1.02469989 -1.86916777
#>  [7] -1.20155131  0.10609602 -2.96796180  1.43232044 -0.80416764 -1.08746381
#> [13]  1.11482073 -2.38346674 -1.38823881  0.05758605 -0.92037899 -0.28523591
#> [19] -3.73837713  0.66580968 -1.21123944 -0.11062152 -1.53211445 -1.79235160
#> [25] -0.96819825 -2.46374849 -1.52182315  0.27783022 -3.07647921 -0.40049846
#> [31]  0.47661620  0.92042331 -0.36969213 -2.04997362 -2.27747705 -1.39467558
#> [37] -2.60328341  0.87199084 -3.49925847 -1.25581180 -1.61353575 -1.65610830
#> [43] -0.02355573 -3.58733104 -1.18061859 -0.86982618  0.57125270 -0.04931671
#> [49] -2.60927107 -0.77098124

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.740626  5.900293  7.125981 13.068378  9.867039  7.380626  9.505898
#>  [8]  8.406207 13.013916 10.620875  6.569293  5.490735 12.819482 14.359600
#> [15] 11.536437  3.696757 11.590383 18.596667 10.821823  8.258019 12.447899
#> [22]  8.789981 17.091986 16.492554  6.440688  9.897381 15.671837 12.898933
#> [29] 13.586875 10.160528  9.676464 11.942428 11.174638 11.174346 10.623955
#> [36]  3.409679  6.607261  7.025718 11.219906  8.070302 10.958722  5.089413
#> [43]  7.040977 11.200424 11.229340  9.967605  4.617204  8.288637 13.330816
#> [50] 16.092893

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   0.4757406  -8.7663444   4.3851493   6.6659639  13.3964225   5.5228075
#>  [7]  -1.4628045   2.1711685   7.9576105  13.7884178  18.9558178  -7.8023595
#> [13]  -3.8949678   6.3968729  -0.8267832  -2.6419753   3.2903759  -4.9258986
#> [19]  13.0008333  -7.3755247 -17.2798768 -12.4536039  12.4415815  -2.5497391
#> [25]  -0.2196382  -0.2405340  -2.9297342   8.3566332  27.6661886  -4.9043634
#> [31]   2.1258679  -4.7725540 -14.5475084   6.7678264  -1.8920310  14.1406906
#> [37]  -4.9710947   1.9289907   9.1304843   8.9803506   8.4233749  -1.6229876
#> [43]   5.0742649  10.4491989  -3.8215767  -2.1237513   6.1308390   0.2907420
#> [49]   8.8163930   1.3690395

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.251171