Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.115752551 -0.015683460 -2.272235620 -1.359175836  2.005572922
#>  [6]  0.652232323  2.150610583 -1.182311786 -0.109387484 -1.771087495
#> [11]  0.827859681  2.963556423 -1.369467907 -0.769744106 -0.058443096
#> [16] -0.027804323  0.595547496  0.057430672 -1.767049266  0.477691512
#> [21]  1.049325473  1.342799188 -0.487119093 -0.616838875 -0.612460805
#> [26]  0.804466276  1.889487836 -1.301994505 -0.918600603 -1.610500411
#> [31]  0.131628326  1.972342981  0.008205664 -2.272311760  1.460327799
#> [36]  1.715940014 -0.586160862 -0.120257492 -0.699148000  0.955806926
#> [41] -1.011514035 -0.849125237  0.153765309 -1.463716054  0.538385689
#> [46] -0.796602654  1.425306472  0.605709729 -0.055078094  0.775507274

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.540910 12.374905  7.169145 12.283602 15.584641  8.979480  9.990270
#>  [8]  7.157298  4.573490 14.456177  6.637715 16.528729  7.687656 10.810428
#> [15] 13.500248 14.150397  7.924722 11.785708  4.846675  5.329103  6.141319
#> [22]  9.848414 11.988030 14.233121 14.260007 11.898250 10.248171  8.537050
#> [29]  7.829430 16.582580  6.429423  8.627046 10.256487 13.923028 12.010429
#> [36] 13.920404 12.250233  9.184488  9.667579 10.571504  6.591944 12.905294
#> [43]  3.090364 11.570282 12.205416  7.623788 11.767522  9.619393 10.550328
#> [50] 15.323183

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   4.12056822 -11.99931745  -5.10610886   5.44605779  11.42773193
#>  [6]  -4.51482669  -7.66489711   7.50638439   9.18953665   4.65756934
#> [11]  15.05410993 -20.92793237  -5.37919997  -0.98017653   7.15242033
#> [16]   2.68240188  -3.05389352  -3.19293464   2.00770906   3.44669120
#> [21]  -2.56090461   3.13020170  -0.39469141  10.32567721  -4.93555292
#> [26]  16.42677380 -17.13526262  -3.02781766  13.78450479  13.27948907
#> [31]   7.76072113  -1.02148547 -13.08426377   4.99919917  13.23205017
#> [36]  -9.38168833   2.57974859   2.55682881  14.53421146  -0.78473965
#> [41]   1.89923140  -7.35959883  -0.03914916  -8.14653496   1.63022433
#> [46]   4.91240665 -23.04536952   0.48385549  -5.31772438  -5.22183848

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.7410381