Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.86002533  1.02513505  1.31114956 -0.08701409 -0.29881141  2.91586477
#>  [7] -2.37834480  0.40473661  1.82244942 -0.16478528  1.09322479 -3.36838433
#> [13]  1.69415703  0.37705228 -0.33099463  2.45204751 -0.32326376  0.06730901
#> [19]  1.34837103 -1.99886192  1.26125491 -0.34136086 -0.79270141 -0.74030068
#> [25] -0.78253923  2.89758535 -2.61652598  1.75781223  0.33657481  0.26263498
#> [31]  2.23529714 -3.03759085  1.82952600  2.35797655 -1.73307770  0.23192774
#> [37]  0.05832130  2.22604579  1.12518783 -0.98607758  1.74355594  0.92747066
#> [43] -1.24871948  1.95892263  1.64723171  1.43292639 -0.78225569  1.94874542
#> [49]  1.75142821  0.12004379

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.976554  6.325031  9.774664 13.333600 13.153573 14.122086  7.524760
#>  [8]  6.787937  9.841033 15.630538 11.701313 10.827239 11.415815 10.334830
#> [15] 13.958185 16.367133  7.835268  8.534228 17.178148 13.069275 16.350209
#> [22] 11.614882 11.536802 17.002345 10.684406 18.358744  4.223902  6.522947
#> [29] 13.332178 11.602828 12.153446 11.975106  8.518737 15.724402 13.649742
#> [36] 11.417936 11.060159 11.863029  7.956172 11.414222 17.209462  4.820306
#> [43]  9.644382 17.212363 13.736767 10.253006  7.486111  7.049639 14.235273
#> [50] 11.126631

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -3.1994251  -7.2141308   4.7560697   9.5975028  -3.5384455  -3.2254150
#>  [7]   3.4211180   2.9211196  19.1366487  12.1821180  -6.0728771  -5.0893523
#> [13]   4.0330090  26.6325659 -16.7825117  -4.3441483   4.8738898   8.8099345
#> [19]  18.4214985  14.9676294 -13.2351831 -17.7004979  -0.3798901  12.7322120
#> [25]  -4.4377000  -9.6301049 -27.2633550   5.4512086  18.1119049 -12.1329462
#> [31]  12.1617762  -7.5691349   7.8441381  13.2037130   4.4569716 -10.5079660
#> [37]  -2.9525023   3.4911835  15.9982516 -10.7088358  -7.8863824  -2.6111438
#> [43]  -4.6681121  20.9378970   6.7753019   3.7817002  -5.3141264   0.9820096
#> [49]  14.3560166   4.1051624

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5117394