Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.3287916644  1.8176367141  2.0640380282 -2.8293116694 -1.3169580715
#>  [6]  0.4911912760  1.3450796853  0.1398358454 -0.7020182146 -1.1258211459
#> [11]  1.1799354771  2.4266861444  0.2635425399 -1.2897431310  0.0006033693
#> [16] -0.6267391884  0.5633798379  0.6443952881 -2.0721109669 -1.8672098248
#> [21]  0.7961151066  0.3146689257  1.1321196383 -1.3288665889  0.2958692525
#> [26] -1.6865063591 -0.5776595211  1.1906446331 -3.4588915916 -1.7244475842
#> [31] -1.4802097026  1.7041908897  3.0557462241 -0.7537059536 -0.0039669793
#> [36] -0.7950929253  2.2031465086  1.9618930268 -1.0668172423  0.5744921744
#> [41] -2.7901717354  1.1984824352  2.4841884305 -1.1302588213 -0.3159954775
#> [46]  1.2438218467 -0.7116612299  0.7666593360 -0.7724028427 -0.7277045526

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 16.095705  9.590178  8.396997  6.901987  8.040784 12.741968 15.481750
#>  [8]  9.735908  6.040899  7.949420  9.692764  9.781296  9.213107  7.839468
#> [15]  8.429815  5.281702 12.037277  7.049507  5.423284  6.790113 11.461113
#> [22] 10.304266 10.085327  7.072763  8.292558 13.456900 12.888281  8.809781
#> [29]  9.607437  8.197775  5.310515 13.695811  5.741901  5.242663  6.031121
#> [36]  4.098346 12.042564 10.539028  8.998762 10.428101 11.713628 12.815709
#> [43]  6.517428 10.355202 10.316373  5.940547  9.689459  8.496178  8.536730
#> [50]  8.134365

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   6.1775214 -15.8972064  13.1982039   3.7000909  15.9083248   3.7751139
#>  [7]  -0.9057073   6.2953289  -0.2404765  -6.0215435  13.8130259  -2.2273346
#> [13] -10.8641314   7.7798589   3.2250666   3.3125341   2.9417800  -5.0143992
#> [19]  18.6427521   4.5512856  -5.0320025   4.4330880  -3.7907824   9.2221613
#> [25]  -5.2906691 -10.5174351  -0.1958789   0.2724060   2.2069266   2.8558876
#> [31]  -7.6213460  -3.3692167   4.2753852   1.1194804  -0.4673722 -19.0347338
#> [37]   2.9364675   8.6259369 -20.6528984   1.5337949   4.1212238  -2.2268260
#> [43]   4.4666644  -7.3915447   7.0435838   2.0682463  -3.8461749   6.4030356
#> [49]  -3.3572532   1.3280442

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3901012