Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.118005313  1.523276468 -2.783825967 -0.301386368 -0.508123915
#>  [6]  0.875949418  1.954416434  0.830305585  0.600895489  0.133621270
#> [11] -0.616037759  0.816334840 -2.190863311  0.150702094 -0.713462594
#> [16]  2.028703692  1.330013666 -4.560680420  0.112788901  0.009896491
#> [21] -0.682847972 -0.004432157 -2.124879241 -0.399062196 -1.520333007
#> [26]  0.118816012  0.201539741 -0.912162104  0.347248429 -0.362792512
#> [31]  1.930330586  1.967610721 -2.078723375  0.479821223 -1.190902208
#> [36]  1.176212254  0.148747750 -1.497261821 -1.390134776 -0.992597578
#> [41]  0.085204886  2.762897486 -0.855539942 -0.959125300 -0.605940028
#> [46] -0.457999740 -0.372622416 -1.856046660 -0.855910180 -0.953720525

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.8709015 15.4560200  8.7113353 10.8143984  9.9495577 13.7255306
#>  [7]  8.6387165 -0.1181029 12.6729312 12.1046195  7.3852167 10.0811722
#> [13]  4.7026657 12.4843571  9.5289253 10.3290312  9.5583195 10.5230161
#> [19] 11.4713484 14.4234824 11.3275768 14.5169510  8.5654564 10.7642541
#> [25]  8.6732411 11.0388115  7.2665726  1.1255332 10.3954988 11.4532453
#> [31] 12.6713295  5.1879943 10.4111730 11.0894401 11.1226339 11.3290401
#> [37] 10.8231937  3.9388823 11.8737758 12.0630343 11.2273509 12.8130915
#> [43]  5.3911797 11.2357787 13.4121451 11.0416909 10.4535462  2.1239461
#> [49] 10.1475429 15.2312034

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -9.5587214  -4.2207773  -2.0993735   3.4308234   2.5656915   3.8047241
#>  [7]   3.4446847 -17.5333383 -18.6742153 -14.1335524  -2.7894112  13.8380030
#> [13] -27.8799410 -13.6162335   0.3341167 -11.9082156  -1.0174441  -9.8137972
#> [19] -23.8949700 -10.7696732   4.5218187  -1.5748473  -3.6338411  10.3987335
#> [25]   4.2071797  12.0785959   1.6252793   4.2400314   3.5220872  -0.2939461
#> [31]   3.4928515   2.9145106  -5.1790576  -7.3752983   2.6766732  -7.4506679
#> [37]   7.2798756  -6.0732740   0.4576653  -7.8511897   5.6935893  -6.2032133
#> [43]  -6.2554892  12.1255094  -7.3556053  13.7499177   7.9301652 -13.5846764
#> [49]  -3.2242597 -16.9037294

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.7367899