Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.17064382  0.59178687  2.28209642 -0.35415096 -0.38583263 -1.08067682
#>  [7]  2.74238369  1.80478535 -2.64813433 -1.48979302 -0.09622916 -0.48555658
#> [13]  0.64042258  0.10950634  0.26530657 -0.24833189  1.36255913  0.97863966
#> [19] -0.94730500 -2.26353570 -0.36696609  0.44955578  2.21947979 -0.68672035
#> [25] -0.38163740 -1.78405736 -1.14319663  2.53813964  1.01961977  0.71722105
#> [31] -2.37466495  2.20399030  1.82005210  1.34737292  1.13077070  0.35197173
#> [37]  1.57727564  2.02790360 -0.11409196 -0.55705546 -1.41983180  2.05123968
#> [43]  1.07994234 -2.46176794  0.69690745 -1.05652092  2.39699020  1.43608647
#> [49] -1.02483607  0.16006291

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.525945 10.199260  9.262392 14.061442 13.002997  9.722890  6.813585
#>  [8] 12.872824 10.498132 17.243937  8.849560  9.911803  8.888829 11.395308
#> [15] 12.496550  7.015007  9.660029  8.663305 15.380054 13.799707  4.422925
#> [22] 12.760369  8.964927  6.254166 16.276008 10.426835 14.677740  8.571456
#> [29] 11.921611 13.856624  3.957067  9.353125 13.935207  8.268327 13.826020
#> [36]  8.808246 12.147306  7.307419 14.287173 17.124290  6.514142 13.308700
#> [43] 11.427631 13.158163 13.124521  5.467556 11.644113  7.946738  5.619816
#> [50] 10.437452

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   7.2736526   0.6321739  -5.7420160  -4.2003122  -3.1575290 -15.2631503
#>  [7]  -1.5533846  -5.6619953  -3.0591386   3.3072728 -24.1853513 -10.5600142
#> [13]  11.2657536 -15.9161362  13.2458046 -23.2968686  -6.2655615 -10.5230338
#> [19]   7.9052058   3.2048166  -7.4732581  -8.2490255 -10.2782366  17.3151288
#> [25]  -0.9783644  -1.6025109 -13.1899907  -0.2043025   1.6667717   4.4483731
#> [31]  -9.2733238   2.2778257   7.5427752  11.9200695  -6.6344581   0.7348605
#> [37]  -5.6603201  -0.1957461  -0.5506914  -3.6419571  -6.8059081 -10.8225552
#> [43] -16.3545114   4.9867788   4.2156642   9.8899737   6.2265813 -16.9358328
#> [49]   8.6245831  -9.1151719

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5758747