Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.08678509 -1.60602232 -0.38477798 -1.36290297  0.10213103 -1.34628357
#>  [7] -1.63731426  0.39369888 -0.79823411  1.43745353 -0.30141397 -0.49401562
#> [13]  2.01705583 -0.74723716 -0.19901225  0.58476580 -2.79518748 -1.25794282
#> [19]  0.34412759  1.78467286 -1.08611728 -0.49481446 -0.00490684  0.06531705
#> [25] -1.68835747  0.25422343 -2.29343869  0.38041871 -0.65745484  1.67441260
#> [31] -2.26199742 -0.74368065  0.27742093 -1.68415759  1.02346689 -2.73021170
#> [37]  0.35646428 -0.59298311  0.60761888  0.78245168 -0.61656420  0.06156292
#> [43] -0.79461296 -1.12996721 -0.72353002 -0.52195820 -2.72692655 -1.81594970
#> [49] -0.74952903  1.52410410

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.660867 10.662776  8.845902  9.451240  5.523393  5.492712 11.108449
#>  [8]  6.113482  7.983535  4.375697 10.595848  9.045481  8.077544  8.177247
#> [15]  4.419949  7.008503  8.063235 13.737089  5.566769  8.520123  9.311700
#> [22] 13.645246 11.247526  7.590579  6.292680  8.895414 12.525637  7.416070
#> [29]  1.337768  7.511058  5.719399  8.611643 12.201664  2.016374  6.866098
#> [36] 11.055502 12.122462  8.862209  7.049424 14.356687  5.901652 14.967035
#> [43] 15.556277  6.660075  8.538771  4.744063  7.014096 11.676174  2.268712
#> [50]  8.406624

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.63463375  -3.17367349   5.20919187  -1.30263524  -0.60820204
#>  [6]  13.33778158  -8.08854859  -4.64358403   3.84714562   9.25288711
#> [11]  16.35998084   0.07089643  -4.71966551  -3.08052803  15.92039840
#> [16]  -2.37294678  -5.97818537   3.11614630   7.06666029   7.12087820
#> [21]  19.85563218  -8.06921147   6.05933240  -4.15363928  14.66602434
#> [26]   5.06774872  -2.48294547 -16.76864764   2.15143183   2.58357006
#> [31]  -1.04012487   5.48544840  -8.61063562   0.52577040  25.11680945
#> [36]  13.04131882 -17.66546255  -6.31399789  11.41773553  -1.17560946
#> [41]  15.96911509  -8.80067020 -11.87530893  -8.21162192  11.17053812
#> [46]  13.30556617  12.14543020  -2.55404338  -4.18637406  18.82671142

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6746786