Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.04636298  0.70459518 -0.29554296  1.02193640 -2.12634695 -0.30101775
#>  [7]  0.27240405  0.05518623  0.08443497 -0.01207853  1.59218791  0.30622325
#> [13]  1.87631919  1.05011059 -2.86721995  0.34540111 -0.54734021  1.32623587
#> [19]  1.22778341 -2.36160284  0.85770537 -0.42454076  0.05786756  0.12373069
#> [25] -1.97189804  1.21537248  0.37282548  0.40373937 -0.28593391 -0.76759190
#> [31]  0.58875803  1.22281263 -0.78411663 -0.32195463 -0.44722474  1.24447146
#> [37]  0.91431302 -0.65202986 -0.60782808 -1.10103756  2.74598315  0.85094764
#> [43] -0.82106297  0.31761118 -0.25534945 -1.16556320  2.47518224 -0.74166544
#> [49]  0.14481289 -1.94267764

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.140897  3.613328 12.170513  5.936829  5.029120  5.544782  6.161357
#>  [8]  7.109919  2.155222 11.773971  9.867197  9.941506  9.307631  6.972526
#> [15]  7.300838  7.464482  2.135054  5.378661 10.470043  9.475447  7.962178
#> [22]  6.234432  8.851260  1.410747  8.330005  6.402301  5.359611 14.987893
#> [29]  7.406064  5.012734  7.940158  2.299873  8.035931  4.879599 12.953484
#> [36]  9.140221 13.055122 10.903292  2.208718 10.159111  7.828341  3.706038
#> [43]  3.601670  7.808602  4.531655  5.219038 11.027473  9.869865 13.616613
#> [50] 10.887258

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   7.8539279  -9.8876537  -7.3438048   8.1155355  -1.5760464   0.7062493
#>  [7]  -7.6041604 -17.6519162  10.6277662  -1.8977983   3.8569891   2.4265966
#> [13]  -3.7102973   1.3286100   4.3504253   2.4030240  -4.3678514   3.8649216
#> [19]   6.4236193   4.6847652   4.4209856   3.3013068 -13.0390424  18.2426587
#> [25]   1.3926854  -2.4411706   1.7389328  -8.6409568  12.9472611   3.0246581
#> [31]  -3.2672654   3.6719848   9.7023539  11.1596626   3.7443107   6.0495741
#> [37]  -2.3321461  10.5974880   1.5006552  15.9602347  10.5091844  -9.6995450
#> [43]  -0.6571808  19.3862773   1.1743902  -1.3330187  -6.1760763  11.0637927
#> [49]   9.2976207  -4.8006815

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1837075