Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.107300759  0.006957816  0.515775451  0.735750653  0.896212405
#>  [6]  0.728420155  0.254088676  0.130220021 -1.091690591  0.133511953
#> [11]  0.147687460 -1.082457609  1.633997855  0.316846531 -0.592245514
#> [16] -0.300845350  0.091200025  0.641189279  1.648058059  1.679707143
#> [21]  1.355604203 -0.554410114  0.193828782  1.008667755  0.468846888
#> [26]  1.097676509  1.265174421  0.495311578  1.250006466  1.313419755
#> [31]  0.738739855  0.477442313  1.192118230  0.066794435  0.421064928
#> [36]  2.966054613 -1.838281607  1.796258665  0.986236667  0.231823144
#> [41]  0.765738900 -0.315807981  0.286569575  1.432713000  1.103175494
#> [46] -0.714855734 -0.431489357  1.732140230  0.952122349 -0.325825939

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.406579 10.812625  9.611693 10.637371 17.268572  7.233950 10.508596
#>  [8]  8.635929  1.077371 14.236536 11.232206 10.773247 10.426383  1.845157
#> [15] 15.107167 14.125324 11.077410  7.185551  4.277614 18.025986 12.171960
#> [22]  8.672911 15.248856  5.467465 14.159256 12.141811 17.648164  9.962264
#> [29]  7.520714 12.855984 15.464777 13.545317  4.412843 13.072375 18.997503
#> [36]  9.580168 15.512368  7.787724 12.945686 12.096080 10.528488  7.943540
#> [43]  8.679343  2.946920 17.943186 12.158336 13.476167 10.521433  5.571868
#> [50] 14.241354

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.9906528  13.0302287  -3.5641690   7.9761552  -4.8924698   5.2134919
#>  [7]   1.1445207   4.4192744  10.2734708 -15.1026341  -4.3614284   2.5190069
#> [13]   1.1688265   9.7440130 -14.8718005   3.8977310  -1.7516274  12.0685015
#> [19]  17.0219452 -15.8222104   6.2932831  -2.5822019  -2.3978793   5.9840783
#> [25] -12.4211914  -8.6830783   3.1738166  -3.3745912  -8.3002007  -6.6639138
#> [31]  -7.2021398   4.0906050  13.4612568  10.8717808 -12.0407624  -0.1776787
#> [37]  19.8465051  18.9911949   2.9118461   6.5754067  -0.9254858  14.4801180
#> [43] -20.2275409  -8.2990190  -2.4288498   4.1338748  10.2658838   9.7942828
#> [49]  -2.7436374  -7.2655874

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.235227