Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.59856954  0.69983152 -0.45449884  0.87975585  0.24243010  1.06786294
#>  [7]  0.53788460  0.91882128 -2.45609460 -1.71379995  0.50541732  1.53878173
#> [13]  0.66535021  0.68340662 -0.79106236 -0.45391934 -0.37493092 -0.42811332
#> [19] -1.40182026 -1.38721034 -0.61882340  0.34421872  1.32687022 -0.06937670
#> [25] -1.30724777  1.65065940  0.61930005  2.16202725 -0.65127789 -0.03205158
#> [31]  0.30551806  1.18781874 -0.47970940 -1.45821563 -0.58277560  0.58433729
#> [37] -0.02302816 -0.01875244 -1.78170178 -1.71567815  0.38409277  0.93149368
#> [43]  1.53188684 -2.31694323 -2.37823780 -0.08867376  0.02227789 -0.70371716
#> [49]  0.45281046 -1.14157448

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 16.488196 10.207919 11.211879 16.085250 13.889967 12.304780  9.517393
#>  [8] 12.146242 11.054883 13.050603 14.039228 11.589397 12.943793 12.537485
#> [15]  7.550530 11.061009 15.237575 12.026200 16.893078  9.383547 11.751243
#> [22] 17.868616  9.844348  4.078308  8.806133 16.839687  7.328084 10.805421
#> [29]  6.699894 11.287846 12.896451  1.627725  8.030796 10.861190 15.076093
#> [36] 17.251491 13.376913 11.265968 12.683346  3.081024 12.690030  7.871000
#> [43]  8.985802 10.894179  7.555341 12.381806 12.320004  9.698766 15.103579
#> [50] 16.231516

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.2959974  -0.9882488  -1.0934973  -1.8454509 -12.4992267   4.4424063
#>  [7]   3.3467021   1.6618999 -10.9789514   3.6379318  -6.7352259   8.0286119
#> [13]   7.1399740 -12.7156801  -3.3335398  -3.4681914   8.1419265   1.9748192
#> [19]  -0.7103134   0.9882328  15.0435259   1.9781856  -6.7513405 -10.9910596
#> [25] -19.0788652   5.3666432  -5.8690175  -2.6176000 -11.2494103 -13.2070119
#> [31]  -7.2645474 -11.3775410  -1.2171446  -4.1186457  -5.3120676  -7.6280153
#> [37]  10.9024850  10.2651572 -14.0143713   3.0616835  -2.5093052   2.3157861
#> [43] -10.4564595 -12.1927318  -1.6680234  -0.2616487   6.4958539  -9.9536858
#> [49] -16.7124205   0.5521731

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1746739