Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.56602522 -1.83761879 -0.57234899  0.02789260 -0.08390832 -0.51195845
#>  [7]  0.72757083 -0.52720648 -0.48995394  0.86319759 -1.57415647 -2.11952016
#> [13] -0.12675244 -1.32799432  1.96443101 -1.67578720 -4.11175335  2.10975373
#> [19] -1.84178712  1.55963874 -1.38994144 -0.30348460  0.13026460 -0.81445590
#> [25]  1.39208923 -1.50471481 -4.76530249 -1.10517522 -2.90829957  2.30483064
#> [31]  0.41907109 -3.04866407  0.34943543 -0.34585315  2.07249315 -1.32226346
#> [37] -1.86257735  0.90645263 -0.23780001  1.18233970 -0.41768389 -2.96081068
#> [43] -0.24777200  1.14806385  1.13903464 -2.04730988  1.06952799  0.16074356
#> [49]  1.61947625  3.76863843

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  5.485399 14.860614 14.332718 16.657848 11.562885  4.943607 15.376600
#>  [8] 10.049197 17.022214 14.460995  5.682379 14.353591 10.570734 17.907677
#> [15]  8.385195  4.255314 14.468050 12.574023 15.899656 15.694263  8.875944
#> [22] 13.385650 14.656607 13.327805  6.912749 -3.395042 15.882070  8.066331
#> [29] 17.365107  7.061165  7.034772 17.563871 12.669682 14.181227 10.436556
#> [36]  1.947394 14.752923 11.202349 22.239955 12.106596  2.081176  7.091448
#> [43]  4.210269 19.308117  4.935345  2.072030 19.942896 15.123847 18.659623
#> [50]  9.073790

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.51086408 -13.01094913   1.84030002 -10.65989759 -10.97599522
#>  [6]   7.75365613   9.14969632   1.69620146 -13.29006075  20.23403941
#> [11] -11.65063628   7.51411177  -4.11084749  -5.04528695   2.02160263
#> [16]  10.87783880   1.19806199 -17.64491937 -13.22924076  12.42727924
#> [21]  11.17693962  -7.37422265 -11.10447331  -6.16308350  11.46486000
#> [26]  16.25494744   7.80940723  -2.59060378  -5.28290847   4.13967184
#> [31]   7.49030193  -1.10784172   3.22102891  -9.13216818   0.02601608
#> [36]   8.60731861   4.32670372   1.73525466  -4.67010165   0.50296353
#> [41]   5.23270149   9.66683742   9.37129545 -17.42891675   2.30083430
#> [46]  19.51072990  11.55023359  18.69589090 -17.66726608   9.01249559

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5248661