Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.1246144759  1.0593190182 -1.5565022188 -0.5496206069  1.3795504026
#>  [6]  2.1525575697  1.2074938736 -2.1761564352  0.7704179841 -0.5866748074
#> [11]  1.2616018972  0.3927842605 -0.1159414638 -0.9802290150  0.9903778580
#> [16]  0.7503157613  0.5357282133 -1.1243880456 -0.7222698792  0.0039627934
#> [21] -0.1967735042  0.5624940732 -1.8947570790 -0.3004287729 -0.5620000012
#> [26]  1.7112207085  0.2668427520 -2.2206521577 -1.4063252562 -0.3212816691
#> [31]  2.7160059400  0.6901593256 -1.7466622046  0.3118179647 -0.0305365008
#> [36]  1.6257446042  1.3698625872 -2.2479128636  1.3108736241 -0.5221474023
#> [41]  0.2852257745 -0.4332664150 -2.2873587651 -0.1575494927 -0.3599399386
#> [46]  0.5714879951  1.9837170577 -1.0531359705 -2.1061914541 -0.0009013107

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.184745  7.721395  3.950016 11.392399  5.271727  8.128709 12.799776
#>  [8] 11.009505 10.503694  7.718462  9.123800  8.236026  4.041836 11.350151
#> [15]  4.860599  8.813680 11.687839  9.321140  7.590183  8.731731  9.525856
#> [22] 10.927519  7.677948 12.257732 14.713914 15.297999  8.505371  8.877656
#> [29] 11.627514  6.991609 10.416284  7.315738  6.783638 10.141159  9.829324
#> [36] 10.731247 10.551889  7.247595  6.001925  8.391529  8.678934  5.835140
#> [43]  7.354320 12.473449  9.265381 12.401682 10.239748 10.881735  7.568282
#> [50]  5.288241

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -13.9220384  -8.4251994   5.1598869  -5.3719536 -16.1043237  11.4981744
#>  [7] -15.3116209  10.0130612  -8.1084141  -2.4045793   8.5652649 -18.9726930
#> [13]  -2.6046925   4.2376195 -17.8655156   9.8989170  -6.7685953   5.8716411
#> [19] -11.4502638   0.2148877   5.7391721   5.1273456  -1.1873418   0.1222813
#> [25]  -9.8258702   4.8731956 -15.0180440  -1.5594581  -5.6460766  -6.5678298
#> [31]   6.0471858  -7.2398574  20.7862305 -20.0560279 -16.7373653   6.5341921
#> [37]  -4.0787236   9.0660278   8.3656205 -16.0431619   4.8029832  -0.9638298
#> [43]  11.3140045  -9.8300280  -6.0667908   1.3074764  -0.8820773  -0.1590164
#> [49]  -2.9052240  -3.4850226

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.365278