Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.223273170 -0.276639242 -0.007420186 -1.630256488  0.091389400
#>  [6] -2.397916388  2.045590292  0.481374521 -0.932061199  2.314874760
#> [11]  1.805100536  0.162855359  1.548685892 -1.670436988  0.432709549
#> [16] -0.991708328 -0.157960128  2.428934946  0.266967161  0.023790485
#> [21]  0.340650817  1.619033983  0.623592684 -1.524633670  1.769210854
#> [26]  0.834170904  0.525222052 -0.750288420 -0.437373588  3.787400440
#> [31] -0.356027701  1.317615242 -1.337717960  0.560910519  0.634792710
#> [36]  0.727646078  1.389186974  1.417526300 -2.417632267  0.616900419
#> [41] -0.115666005  0.377330488 -0.036793079 -3.038998786  1.357718494
#> [46]  1.859518906  2.330293024 -0.312144589 -0.503787780  0.932314159

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.353287  4.900062  7.253962 13.960106  9.749602 13.621676  6.209354
#>  [8]  4.702812  7.317956  5.022452 11.853546  7.309598 10.353981  7.990590
#> [15] 13.661007 11.296524  9.931651 13.099189  8.375519 12.567255 16.129674
#> [22] 14.585482 12.409249  9.022621 10.273976 10.365920 10.087783  7.705884
#> [29]  5.658946  6.510897  6.833129  5.530901  8.593525  7.135381 12.374985
#> [36] 15.532978  5.555078  9.198086 12.726707  9.720483 18.897454  2.681359
#> [43] 11.717340  7.915974  8.697752 11.193445  3.526015 10.173726 14.523925
#> [50] 10.900106

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.9873969   2.5992218   5.5381791  -7.9174252  -7.2081316  -9.5314498
#>  [7] -13.7098041   0.1422900  10.1665141   6.1042047   1.1679980   0.5184415
#> [13]   5.7412178  -4.0682915  -7.2880253 -14.4691166  11.1909303   2.8337346
#> [19] -14.5103300   4.5082767  -4.2484944   0.5783514  18.9453748  -4.0268229
#> [25]   2.5698472   0.7626729  -5.0034632  16.4916514   3.3122081  -3.2313314
#> [31]   8.1709000   7.5977542   5.4676207   8.7808074  12.5687911  -3.9990585
#> [37]   3.9087468  12.6188709   2.4044570   0.3150423  -1.4388682 -11.8913622
#> [43]  -8.0221836   0.3114869   4.8181837   3.3688141  14.0196366  -1.8959862
#> [49]   0.2781979 -17.2646658

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3438465