Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.16114873 -0.58294925  1.61173720  2.99194323 -0.87583986  0.69080689
#>  [7] -1.49170726  1.51721259  2.89223944  2.78905407 -0.06596224 -0.33594780
#> [13] -0.10213330  1.51471413  1.98578943  1.91933510 -0.48166831  0.26030670
#> [19]  2.28055128 -0.04750637  0.61254153  0.50063429  1.03926212  3.02290848
#> [25] -0.61396453  1.28666739 -0.25794638 -0.23618038  1.77833354  0.30737981
#> [31]  0.25968265 -0.46813777  1.01785557  0.10877187 -0.17956038 -1.82598746
#> [37] -2.42224730 -1.98062086  4.17411291  0.15311812  0.57585371 -1.79559052
#> [43]  0.47685937  2.32457396  1.81659629  0.41173205 -0.17235792  0.94692633
#> [49]  3.92316856  0.12240372

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.721603 10.765256  7.944163 12.753200 10.381537  9.234061 -1.036091
#>  [8]  7.012547 12.905557  9.972451  7.079838  3.800910 11.734898 15.418982
#> [15]  5.140636  5.472893  5.860580 12.398799 10.895574  6.119033  8.429172
#> [22]  2.763634  7.707531 11.086036 12.456706 10.082757 12.562126 13.106606
#> [29]  9.070905  5.660494 13.170779  9.039815  9.844703 13.491866  1.708429
#> [36]  9.069891  1.798658  9.387249 13.051715  8.573565 14.939641  3.538618
#> [43] 10.979196 16.450744 10.657840  6.695777  6.612438 15.443148 13.350946
#> [50]  5.537686

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   0.82105407   4.10166984 -16.18894894  -3.67139986 -15.57212715
#>  [6]  19.15347070   8.30916183  -5.96388986   0.83830725  -1.76259557
#> [11]   3.81705232  -6.23429664   5.85271299  14.97163765   8.78941225
#> [16]  -4.68270184  -5.15723980  -1.99805447  -8.83435430  -2.18563436
#> [21] -12.19009383  -1.94188867   6.07996705  -0.04630074  -7.91239552
#> [26]   2.02023741  11.95240731   1.81566408  12.68110401   3.97559650
#> [31]  -7.57690148  -6.70528523  -1.75547439   4.82415456   0.53147136
#> [36]   1.43617001   6.02340003  -7.29577666  14.70737759 -12.48420799
#> [41]  -3.93647390  -9.08913174  17.92897339  -8.09969606  -9.82803255
#> [46]   4.10725862  -3.22274163 -12.74260194  -1.18092885 -18.36797258

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6051784