Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.33040579 -1.25540326  2.11987431  0.67666664 -1.44266061 -1.47860546
#>  [7] -1.16909331  0.37280295  0.28300344  0.63259019 -0.35972587 -0.40840342
#> [13]  1.34659936  0.17017277  0.99990577 -2.69822733 -0.56665191  1.44755264
#> [19]  1.63353222  0.53048625 -0.55810998  0.59800741  1.75637680  0.08636959
#> [25]  0.61598704 -0.48810427 -1.16272535  1.07181051  1.25014605  0.90034327
#> [31] -2.09423547 -0.06648367  0.41061102  1.33150387  0.40563834 -0.92233433
#> [37] -0.64707548  2.06984938 -0.02125517  0.37302688 -0.53258072 -1.57903296
#> [43] -0.08979308 -0.44479285 -0.90219842 -0.96637058 -3.25269649 -0.09144279
#> [49] -0.67883429  0.19180056

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.142396  7.322919 18.087099  6.772920  9.540202 14.360741  6.388436
#>  [8]  8.884794  8.795832  8.911589 13.174044  9.626051  9.513567  8.886241
#> [15]  9.928508 14.098821  4.244929 11.244712 11.344338  6.840537 12.856651
#> [22]  9.917796  8.342847 -1.032811  8.302108 14.814489  3.686168  3.646063
#> [29]  8.368028  8.757775 14.977583  1.561056 10.448116 10.899899  6.637169
#> [36] 11.209973  9.740854 12.862500  8.717299 13.889841 12.640446  9.527894
#> [43] 10.391152 11.545945 10.214143 12.920947  6.880525 10.229648  9.329147
#> [50]  9.990207

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -18.10097972   8.95158605  -5.32778543  -5.70779925  20.60169732
#>  [6]  -3.01059144   9.36783547  -4.74692189  13.76389578  18.22508266
#> [11]   8.51880244  -1.52575667  -2.24129908   4.20778483  16.09457077
#> [16]  -6.37500177  -3.31703885   8.22385545  -0.03454483  -0.54923937
#> [21]  -5.98463487  -5.28617059  -3.54096178  -0.20898121   2.25266078
#> [26]   4.28160155  10.26594228  -7.86111192  -1.02753302   3.19156508
#> [31]   9.31288549  22.26717441  -0.60058023 -17.75374061   6.82881000
#> [36]   2.23065832   8.01332788 -14.31916453  13.34865108   2.61543350
#> [41]  10.85354439   6.77748732   6.21122581   3.38844687  -4.59101839
#> [46]   1.48384272  17.47324221   1.65413452 -10.80649153   4.00498475

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5260044