Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.40346618 -1.41404965  1.36684114 -0.80077808 -2.03637782  0.73199107
#>  [7] -3.41830651  0.61192519 -1.34299498 -0.75300395  2.50984494 -2.71313944
#> [13]  1.41577179  0.61363553 -0.39987878  0.07493961 -1.40108210 -0.17836352
#> [19] -0.71034735 -1.60281242 -0.82792243 -3.14681536  0.97322667 -1.93155073
#> [25] -3.06728895  0.94286884 -1.33507141  2.76017367 -0.99892498 -0.87790139
#> [31] -1.04758284 -3.76336049  2.06522197 -0.32923371 -0.37196077  1.22720287
#> [37] -3.21488604  1.32989090 -0.36797785 -1.52603069 -0.13575345 -2.96751443
#> [43] -0.10401737 -0.51739022 -0.31768047 -1.31829959 -3.57072858  2.67305185
#> [49] -1.27052170  0.43491043

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.727029 11.972271 13.599219 14.764840 13.108877  6.566291  8.289659
#>  [8] 13.405123  6.190791 10.159181 10.777852 13.560200 14.813048  8.977037
#> [15] 15.454228 13.064540 10.256206 14.429851 11.703371  9.825398  7.303613
#> [22] 10.963946 13.061264 15.420964 11.045743  7.539523 13.795198  7.709478
#> [29] 14.959711  9.029456  6.348252  9.844996  7.128761  7.294206 12.579107
#> [36] 10.878814 12.163189 11.340853 11.777489 12.365786 10.003135 10.461219
#> [43] 16.528529 11.333239  9.418131 12.198864  9.276277 14.882726  6.196085
#> [50] 12.523621

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   8.109745   9.289521  14.791870  10.764068 -17.390445   1.118077
#>  [7]   8.638842   8.260957  16.530913   6.182532   8.897026  -9.208501
#> [13]   5.747852  14.623335   1.216275   2.009622  -2.265500 -10.107552
#> [19]  -5.692387  -6.547024  -3.943970  -7.647545   0.590329  14.918342
#> [25]   2.518414  -9.221259  -4.034405   6.709506   4.252947   8.350036
#> [31]   2.955260   9.267064 -11.947497   6.441424  -5.445937  -9.259146
#> [37]  -6.424872  12.486353  11.214358  -1.239290  -6.856322   1.021277
#> [43]  -2.067456  17.021354  -2.156230  -2.828031  13.271982  -2.881414
#> [49]   6.084157   2.516773

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.679263