Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.68380698 -1.12255377 -0.29490726 -0.62446905  1.98974411 -0.13443055
#>  [7] -2.94232806  0.89303418  0.35844378  0.33624016  1.53425416 -1.20382246
#> [13]  1.26665750  0.87265041  1.01861055 -0.52247286 -2.03856327 -0.07699156
#> [19] -0.30034791  1.89978353 -1.00964458 -0.82183367  0.08443318 -2.47600863
#> [25]  0.91106146 -0.73697470 -1.55230150 -1.30769046  0.67131174  4.00659888
#> [31]  0.47680455 -1.83379756  1.14471202  1.06938559  1.35214116 -0.37542863
#> [37] -1.59768896  0.30272185 -0.24904518  2.14558263 -1.78873367 -0.88819582
#> [43]  3.50345022 -0.35421498  0.94232121 -1.24013574 -1.23757313 -0.10396770
#> [49]  0.89833942  1.95862259

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 15.183898 11.207199 14.550237 16.596150 11.543506  1.976346 12.446102
#>  [8] 16.089061 11.688099  8.559517  6.163350  4.339320 12.513799 13.139495
#> [15]  5.951430  9.551832  9.240330  9.249798 15.737911  9.248359 11.122545
#> [22] 12.313831 15.925737 14.528935 10.993419 12.757472  7.323743 14.863289
#> [29] 11.579127 13.010375  7.566202  9.466922 11.897486 15.099558  7.861612
#> [36]  8.820393 14.590417 12.964778 16.916819  9.669692 12.210078 15.326446
#> [43] 13.996299 16.441224  8.942066 12.151388 10.648199 13.706102 14.906553
#> [50] 11.570507

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -2.2835842  -9.9485435 -13.7621521 -11.1097162  10.8629429  13.2788601
#>  [7]  -2.5484359  -2.0532008  -2.4181001  12.8610699  -5.8413136   3.6078847
#> [13]   0.5019934   1.2532302  13.4029344 -15.1872058  -0.9894525  13.6354608
#> [19]  -6.4312228   7.9463973 -10.0823629   6.1272199 -10.6510475   7.5627099
#> [25]  13.2704461  -5.4269053  14.0773778  10.7625416  -5.9421998   4.3962872
#> [31] -14.8974000   7.6450910   1.7720964 -12.7529989   2.0477262  -5.2462084
#> [37]   5.8495486   2.0270557  -5.4645634   9.5277201   2.5735231  -8.6802522
#> [43]  17.8694195  -4.4383116   7.5774028  -6.7425614   4.0475734  -1.7451872
#> [49]  -1.6848146   0.3494404

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.6698432