Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.11731646  0.74064523  0.69897913  0.35424893 -1.63022359  1.06109431
#>  [7]  0.28978126  2.03168568  0.90284454  1.02396242 -0.03462119 -0.19782577
#> [13] -0.37272785  1.08723042 -1.00383130  0.82861786 -0.08672995 -0.70356911
#> [19] -1.23737959 -1.53936918  0.84573566 -0.55579946  1.63071963 -0.23851864
#> [25] -2.32649366  0.38390413  1.12622823 -0.64229931 -0.16113237 -1.46396774
#> [31]  2.13114757  0.59912946  1.67789947 -1.87862000 -1.40239361 -0.07416314
#> [37]  1.27115453  3.54664190  2.39754692 -0.11516624  1.64883109  0.56813273
#> [43]  1.32030645 -1.07337743 -1.80692744  0.09225451  0.89545942 -0.20474596
#> [49]  0.74584859 -1.75733782

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.345739 10.320848  8.816318  8.766601  9.431842  6.791056  8.189046
#>  [8] 10.556237 11.131215  5.409852 10.464672 15.004158  6.877943  8.970247
#> [15]  7.055534 17.279338 17.520951 12.249867  8.766397 12.709991 11.646299
#> [22]  8.701616  8.134999 11.767018  8.250148  8.965064 12.278474  9.009926
#> [29] 10.496740  7.825983  9.368934 17.157008  3.736339  8.932182  8.123866
#> [36]  9.544921  7.622796  6.679487  8.335803  8.086321 18.050567  9.446111
#> [43]  7.426330 13.420225  5.599517 19.183061 12.559592  5.513961  7.978835
#> [50] 12.365227

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   0.565582084  -5.114539650  -0.915400532  -7.855549872   6.628207980
#>  [6]   6.104820892 -11.655882861   4.364991020  -9.323562370  -7.571370461
#> [11] -15.180962260   7.464269524  -1.398644572  -0.155316812  -8.870342849
#> [16]   6.272811657  -9.996199202   7.360793420  17.579079514   9.078413789
#> [21]   0.068851945   3.416357370   4.724015415   1.900366910   0.417454736
#> [26]  -6.385770902  -7.389871667  18.384835621  -0.007526608   1.989502887
#> [31]  -4.226846032  -3.236629990   7.995434570  -3.359215718  11.542373788
#> [36]   8.880954196   5.652838322  -2.044687176  11.697700364  -6.076398591
#> [41]   6.675645716   5.937880124  11.650543465  -2.414886227 -10.582555973
#> [46]   0.740460376  13.281726012   5.886245753  -5.901738056   9.262364499

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4858007