Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.033035502 -0.710376259  1.471514688 -0.753586214  0.050319694
#>  [6]  1.327876381  1.251951986 -0.025494080  0.366055015  0.592419226
#> [11]  1.644048411 -0.706196673  0.072691527 -0.886960365  0.398359579
#> [16]  1.517089358 -2.013060002  2.081441344 -0.209574191  1.032459797
#> [21]  0.161573164 -0.399924483 -1.989573453  0.467868581  0.567464831
#> [26]  0.004966182 -0.748923556 -0.775786679 -1.194142684 -2.491082477
#> [31] -1.389589956  0.455357617 -0.888205889 -0.797009795 -0.668885455
#> [36]  1.688387029  1.211698440 -0.564890161  0.229717415  0.324598260
#> [41] -0.461128001 -1.832495891 -0.647946413 -1.094698177  0.884735119
#> [46]  0.539316196 -1.558956451 -0.701910812  0.681335635 -0.956356312

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.568982 19.184920 10.233066 11.221997  9.751449  4.935985 15.710259
#>  [8]  5.846028 13.412702 13.130768 10.842895 13.347867  9.108548  7.780113
#> [15]  2.793568  7.326337 14.147928 12.702842 14.179246 10.531601  6.736770
#> [22] 14.186503  6.252883 11.377247  8.825776  6.360521  6.682410  4.337405
#> [29]  9.830001 10.822960  5.512735 14.599580  9.132453 11.110876  9.069600
#> [36]  9.503654 16.533853  9.719478  8.618509  8.413744 12.778223 12.290620
#> [43]  6.444005 17.792499  9.366898  8.125359  9.638988  8.109773 14.602450
#> [50]  7.837425

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   6.8319731 -14.2887197  -3.3443290  11.0178351   7.8514108   1.3252023
#>  [7]   3.7556240  -1.8983249  -3.6182718   6.3960697   5.1704762  -5.7059169
#> [13]   4.4229328   4.4127538  -2.2939131   5.5682121  -6.0285318  -6.1930459
#> [19]   2.7413880   5.0226212  14.0937033 -12.0905153  11.6002978   1.2704921
#> [25]   1.1240381  10.9665328   0.2474014 -12.4335635  -5.7867189   7.0255820
#> [31]  11.6512452  -7.7224285   8.6303489   7.7317124  -1.9136544   2.8130163
#> [37]   0.3390720   7.8482622   2.0609892  14.3264260   5.3859600  -3.5196641
#> [43]  -3.8075627 -15.4280703   5.8581878   2.8268787  -5.1259184  -5.5351068
#> [49]   1.6245680  -0.9409805

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3775199