Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.14821473 -0.34291607  2.18978508  0.50835642  0.12023117 -1.10600246
#>  [7]  0.69208769  1.30415609  0.96512411  0.01229349 -0.05926837  0.38589460
#> [13]  2.80753410 -0.07543410  0.94790394 -0.32900648  1.56622336 -1.40797409
#> [19]  0.29190256  1.06226748 -0.19186452  0.34449890  0.13790193  0.76558655
#> [25]  1.56917328 -1.43335305  1.40858156  1.70508210 -0.48240556 -0.09651219
#> [31]  1.22212640 -1.29090891  1.06780813 -0.80303733 -0.01313358 -0.17578425
#> [37] -0.23716523  1.23610532 -1.88798124  0.31694409  0.12282416  1.53680196
#> [43]  0.83218124  0.52530986  0.72667200 -1.56407326  1.24459714 -0.75690047
#> [49]  1.92377883  0.58749695

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  2.5965093 14.7410788  7.1221126 11.3482713 11.5059022  7.5029279
#>  [7] 18.1375811  5.6121842  9.5393616  9.3300498 13.6927205 18.1158961
#> [13]  9.4919645  4.3159992 12.3574785  6.6506849  6.8355084 -0.6793377
#> [19]  9.6353050 15.5772172 13.0781536 20.0094513  9.1261170 12.2109605
#> [25] 10.7401818  4.2427095 10.8214797  7.4940515  8.1687738 13.0068735
#> [31]  4.4911809 11.8825205 12.2781462  2.3424504  7.5163683 12.0896702
#> [37] 13.4399725  3.4227629  7.7793800 12.1147190 12.7979100 11.3636052
#> [43]  4.5449008  8.4733306  7.8687580  8.7432532 20.2901125  5.2761236
#> [49]  9.8961503 12.9779518

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.50274261   4.68219279   0.04797097   4.87219126  10.77288675
#>  [6]  -6.79661972  15.46663265 -16.15838659  -2.74236405   6.14239797
#> [11]  -5.08882593   2.80269721   1.76908722   6.48013201  -0.89199381
#> [16]  10.55066379   4.38619763   4.22463201  -1.62969654   5.14472675
#> [21]   0.40864732   5.97960678  -6.43688590  -6.44292811   4.07066043
#> [26]   1.15733034   6.69928647  10.97734544   0.99834296  -8.82699254
#> [31]   1.86391212   7.43518155  14.23461453   0.44305748   8.16476978
#> [36]   0.75029858  14.08615469   2.15455625 -13.33345442   0.27529256
#> [41]   7.85292476   8.65166988 -12.34562127  11.03607210   1.33482823
#> [46]  -1.68222742   5.36340412  15.67458779  -8.24850024  12.55398527

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5308067