Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.65575131 -0.76130210  0.27499057 -1.45337245  0.07271368 -1.50828659
#>  [7] -1.57899724  2.04139563 -1.75525691  0.92236950 -1.47086145  1.19834934
#> [13] -0.63915522 -1.04340375 -1.20757944 -2.70785404 -3.55965660  0.17896765
#> [19] -1.90932096 -0.28506630 -1.46880221  0.49951868  0.50615323 -1.81865854
#> [25] -1.32805902 -1.81093141  0.79231317 -1.71635787 -1.95606222 -0.06898579
#> [31] -1.99308361 -1.23831966  0.52208021 -0.72855296  1.19187859  0.18475271
#> [37] -1.14300227  0.85899278 -0.34888315 -0.94854568 -3.32826928 -0.59792234
#> [43] -0.66011147 -1.00246301  0.07011579 -2.92870139  0.34270911  1.70917937
#> [49] -0.67905846  1.20322808

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.653856 13.751838 14.465749  9.729029  8.256555  3.794937 12.666979
#>  [8] 10.565988  8.646719 12.501644  7.055214  8.194272 14.821624 11.479515
#> [15]  7.154386  7.778118 12.266046 10.032156 11.959645 12.072922  7.943790
#> [22] 16.157448 10.957016 10.091318  6.666688 13.440426 12.295015 17.067310
#> [29]  6.166485 15.818077  2.898347  9.496249  8.188356  7.411519  8.013714
#> [36]  6.356617  6.997180  9.515486 11.787060  8.053815  7.343309  9.143882
#> [43] 13.269961 13.105339 11.034362  8.701936  3.720933  8.892263  7.644863
#> [50]  6.617496

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   2.434820547  -1.410260300  -9.398730450   4.291168174  11.311397010
#>  [6]   4.348471142   3.313425078  -0.817610302  -0.012669485   2.125479114
#> [11]   8.373645189   6.803956462  -2.982191199   5.869498330  -2.997731657
#> [16]  10.367883150  -5.418164556   1.136630219   4.847021135   7.939098229
#> [21]   9.798598265  -5.999127074   2.514809609  10.782644772  -2.651930903
#> [26]   3.449059163  -0.002499503 -19.141523701  -1.218053128   6.398107830
#> [31]  15.788662714   0.517258559  -4.332879240   5.489984698  -7.746607995
#> [36]  14.059748295 -15.511877733 -13.994950038  16.259589137  -3.980764836
#> [41]  21.228171124 -10.669687380  -7.793123426   7.025438983 -13.150060789
#> [46]  -2.672771467  -4.128440907 -12.466243219   8.469804224  -1.957353185

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4443458