Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.73086164 -0.38793636 -0.23895576 -1.93380723  0.74155568 -0.49782648
#>  [7]  0.54595517 -0.08538860 -2.13214448 -2.61191596  1.53919822 -0.38936690
#> [13] -2.88314106 -2.27641237  0.43763366  1.55914667  3.42422617 -2.28623918
#> [19] -1.45486195 -0.69316902 -1.04882020  2.44729929  0.39117327  0.66813606
#> [25]  0.25188195  0.85995808 -0.62980406 -2.38819289 -1.58014570 -0.01585593
#> [31]  1.12942598  0.12618952 -0.69337782 -2.09816707  1.07487950  1.40749450
#> [37]  0.29184077 -2.29796653  0.57745785 -0.07550786 -0.46817306  1.49136564
#> [43] -0.39102403 -2.58160592  0.73854934 -0.83366626  0.73872332 -1.75751393
#> [49] -0.94645677  0.81434710

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.485242 10.881083 13.766543  8.371173 13.699388  8.607344  8.641258
#>  [8] 14.121460 12.618679  9.864017 14.902792  1.003609  8.831580  8.841253
#> [15]  8.515224  2.158083  9.282160 12.726069 13.420477 10.597962  4.499824
#> [22]  5.764227 12.994147 10.502550 19.079072 11.710679 10.212917  9.766723
#> [29]  5.470045 12.521222 15.757025 10.665290  8.027133 13.059554  7.321083
#> [36]  9.791925  5.466834 16.914355  6.523379  9.438834 17.177652  8.263451
#> [43] 11.844888  9.887888  8.388869 11.217650  8.187658 14.078463  5.284170
#> [50] 13.657134

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -8.5459601 -18.3673880  -3.9603150  14.9142517 -11.6464527  -0.9227335
#>  [7] -12.6937135  -9.3481280  17.9965921 -19.8110555   0.7556251 -20.5757968
#> [13]  10.3542326  -3.5672749 -15.2457962   7.9178795   4.5011154  -5.0526260
#> [19]  14.8170410  -4.6695453  14.3362965  -0.8839721  -0.4631304   8.1995956
#> [25] -28.2389688   0.6703872 -11.7628110  12.2488866  13.3914784  -0.7770188
#> [31]   6.7448429 -11.0327546  -9.5590365   5.6177049 -29.5482036   6.8871166
#> [37] -11.8264970  -5.7735353  -9.3709895 -11.1819696  15.3012854  -9.1473881
#> [43]  -3.0445053   3.9442326  -6.1697397  -0.3557631 -10.4086097  21.5839200
#> [49]  12.5659357 -16.0710447

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1695709