Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.900663036 -0.493234891 -0.536872719  1.391104470  0.935985502
#>  [6] -1.113726970 -3.335388117 -1.695746335  2.306852198  0.384461363
#> [11] -4.709129966 -2.081363560 -0.763863534  0.301345294  1.685044461
#> [16] -1.735366736 -0.009465843 -2.125228980  0.071038577 -0.262607472
#> [21] -2.176388325 -3.325253157  0.328698951 -0.583122082  2.792550075
#> [26] -2.113563273  0.696908102 -1.069453329  0.634308265 -0.893155658
#> [31] -1.698312700 -2.384595537 -1.196280820 -0.564511819 -0.248215246
#> [36] -1.267425127 -0.758069377 -1.513728826 -0.156716254  1.500845903
#> [41] -2.353444282 -1.632409564 -0.648306664  1.916151857  2.743053516
#> [46] -3.203015866 -0.938750498 -1.367344235  0.295515555  1.504689742

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.848415  8.972306 11.973520 15.253162 10.533485  6.943566  7.811935
#>  [8] 10.384894 11.160482 12.408945 12.291538 11.965986  2.582096  6.619067
#> [15] 10.792425 13.859738 14.650333  8.266259 13.059761  9.613750 11.453985
#> [22]  9.443766  5.782673 11.553200 14.083572 17.037595  9.137070  5.529864
#> [29] 11.572172 11.439727 16.180342  7.674755  5.668465 15.559024  5.781788
#> [36] 12.246009 12.985041 10.611202 13.904048 14.830506 13.151398 11.241582
#> [43]  2.671364  8.236953  6.455904 11.782539  8.548688  7.019951 12.120631
#> [50] 17.276700

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   9.9018186  -3.1087451  -8.4128613   9.3481711   2.4517940   5.0010589
#>  [7]  -5.1928385  -2.5347566  -9.9836744  -4.9377584  -3.2811941  -3.9693551
#> [13]   1.2013766   3.7727486   0.4488928  -1.0943496 -11.6469171  -8.1433981
#> [19]   7.3933687 -19.5522849   3.0956822  -2.5411120   1.4099203 -13.0050897
#> [25]  16.7663273   2.3583051   2.6623186   1.3811056  -3.8478154   9.6675187
#> [31]  -4.3772640  -4.8461016   3.2181389  -1.5316644 -13.4738568  -1.3026575
#> [37] -18.5075551   2.5816893   5.0538632  -6.1673435  -4.1952306   5.5926336
#> [43]   2.0807818 -17.2505310  -1.7777695   2.7005263  -5.4694432  -7.8916740
#> [49]   1.1804399   7.3456168

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.5198895