Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.49450296  1.76326530 -0.55723878  0.08719267 -0.58982440  0.41187134
#>  [7]  0.31681196 -2.48155338 -2.05668730  0.48163113  0.67683591  0.50503756
#> [13] -3.44068085 -1.58009973 -1.99125551  0.87100508  0.63747316 -1.10026840
#> [19]  0.91781115 -0.22907482 -0.03180804  2.44883539  1.54599058 -1.59685492
#> [25] -1.13602622  0.46173735  1.21439851 -2.15201588 -1.39765489 -1.10040712
#> [31]  0.78269256 -1.73290466 -1.24014900 -0.05607784  0.35269798  0.07022130
#> [37] -0.86342067 -0.33794203 -2.14468846  2.90714464  0.44008625  0.82500833
#> [43] -0.72053203 -0.06855008 -0.27195064  0.46892913  0.63223643 -2.73495496
#> [49] -0.67569203  0.21638484

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.992799  5.836268  9.646364 15.712887  6.046423  4.528281 18.739017
#>  [8] 11.319318 14.432769 14.424811  7.875074  1.075277 11.836508  4.062464
#> [15] 13.959939 10.290885  6.737884 11.662991 10.001451 10.530494 13.955500
#> [22]  7.654693 11.135830  8.791852 13.262444  5.079954  8.483116  8.023099
#> [29] 13.643348 13.184658  8.847046  7.308258 11.910012 15.368235  4.974050
#> [36] 13.123873  8.798924  8.177686 11.880163 12.124097  2.255502  9.853272
#> [43]  7.899634 13.256797 15.299892  9.667233  9.858879 12.275391  5.719877
#> [50] 11.042171

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -7.0161242   2.2637241  -6.0678605  13.0269489 -12.9155057  -4.6323822
#>  [7]  12.6146813  -2.2860220 -15.7113355  -3.5728355  -4.2883575  -7.3617796
#> [13]  11.0031561  -9.8653074   3.9161756  -1.2014407   8.1175695  -0.7503631
#> [19] -13.0781979  -1.2563850 -18.0520004  10.5115877  -5.4695698 -10.0045940
#> [25]   7.5232486  -3.9016369  12.4386966  20.5090791 -15.5414033   9.0884797
#> [31]  -8.4246907  24.0667158  -0.8491765 -17.2169570 -11.4263821  -2.4347574
#> [37]   7.6044161   5.8371619  -0.5230782   8.7462005  -9.9910095   7.9373810
#> [43]   4.6591841   1.1189162  11.4303420   1.3044758  -0.2229242   2.4042641
#> [49]  -4.5025304   9.0638419

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.2097463