Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.005190511 -0.993000747  0.259556871 -1.735405646 -0.733560653
#>  [6]  1.200983899 -0.233833311  1.448374737  0.017633545 -1.620845967
#> [11]  1.541416252  0.389605761  0.278635627 -0.936207926 -0.857109118
#> [16]  2.393350492 -0.645971586  0.107360866 -1.793384410  1.314835003
#> [21]  1.462535875  0.098461615 -0.511291262 -2.323183074 -1.085958009
#> [26] -0.062165624  0.843596172  1.052242863 -2.414452735  0.246708393
#> [31]  1.928388296  0.007546782  1.427060373 -2.809617662  1.855267709
#> [36]  1.253503589  0.578907937 -1.033152464 -1.938620504  0.412465436
#> [41]  0.642120972  2.429270575  0.863253817 -3.013820132  1.724393107
#> [46]  1.252022776 -1.160377565  0.554081425 -3.454762088 -2.388840426

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.572424  5.272328 12.742828  8.619264 17.122696  9.752955  4.377966
#>  [8] 12.882681 12.239991  9.913315  9.154163  5.458654 11.436135  6.190182
#> [15] 16.214734 12.085139 13.507190 14.096829 12.401178 12.473911  9.568873
#> [22]  9.539370 21.285277  7.349248 16.864677  6.252736  7.061859 15.171175
#> [29] 10.962188 10.749417 11.915328  8.470118  7.480627  7.216909  9.806216
#> [36]  8.764285 11.928274 15.752171  9.858120  8.613124 10.583482 10.708864
#> [43] 11.655876  3.477654 15.535817  6.314733  8.158466 11.025226 12.116582
#> [50]  7.536193

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -3.5233388  -1.1415069  -3.0687022   0.2221482   9.6703174   9.6622173
#>  [7]  -8.1350269   1.1007334   9.7230293  -8.3210017  -9.6156234 -17.7859998
#> [13]   3.2510895   6.3703749  11.6574636   4.9191244 -14.7029286  -2.3432604
#> [19]   4.5467242 -13.7387768   0.9124925  -7.2535107  -2.9945757   9.7457032
#> [25]  -5.8563539  -5.4396624 -19.0603341  -9.7038322  -6.5489915  -3.4211899
#> [31]  -0.2979641  -0.2995697  -9.7315416  -0.8799827 -16.7656221  -2.9879336
#> [37]  -0.6773888 -29.8890752   4.1300767  -0.2009920 -11.8639184 -29.6728065
#> [43]  -6.3209177  -7.2981018   2.0914048   2.2899612   1.3283245  -1.9790393
#> [49]   3.6249704   2.4190080

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.08053864