Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.55398916  0.09115945  0.89891974 -2.51113583  1.80951176 -0.86613912
#>  [7] -1.37907333  0.34246881 -1.37335414  1.00134129 -0.47217814 -0.87132669
#> [13] -0.24717278 -1.97684041 -1.36414243 -3.27466684  1.28049175  0.53952500
#> [19] -1.09929621  0.47459044 -1.05465532  0.33340217  2.62059717 -2.70870375
#> [25]  2.01848525 -1.02212223 -0.29777071  1.35773438 -0.15001006  2.60973698
#> [31] -1.03629997  0.89236646  0.04499019 -0.84539666 -0.36057076 -1.47797273
#> [37]  0.51335301 -0.90456856 -0.33438494  0.94107722 -2.04682254  0.43414635
#> [43] -0.23384083  0.28249373 -0.45425794 -1.97272706 -0.20872566  0.81140314
#> [49] -2.25911891  0.54379594

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 14.4128817 13.5217489  6.7641645  2.1376453 12.8187831 13.2017341
#>  [7]  9.9969801  3.0854483  9.5446340 10.9874065 18.0083336 13.2884653
#> [13]  8.8790831  6.5682741  9.0113913 13.2337450 13.5721143 13.6799632
#> [19] 12.0124160 13.5953029 16.3394896 11.5906269  9.6254541 13.7626298
#> [25] 18.4437900 16.5226535  8.2293419  8.7129575 15.4173618 12.3199833
#> [31] 13.9148647  3.9648360 11.6638662  0.9815527 18.1507792 15.1531623
#> [37] 11.1763577  7.6832008 13.5271667 13.2867479 12.8968922 14.3045559
#> [43] 11.2616948 10.6501149 16.1697250 12.1640511 12.9798281  8.7089371
#> [49] 14.7273159  5.5673047

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  20.9682938  -9.8499501 -15.1589253  10.6927001   1.1835506  -2.2006761
#>  [7] -11.0067139  -1.9224352   4.9536258  -4.5278945  -1.5931498   2.6024306
#> [13]  -6.3080332   9.6495046   3.7934715  -2.8798353  -9.5024546  -3.7729906
#> [19]  -6.0131406   0.3118065  -2.6143435  -9.2620557  -6.6653992  -2.6263961
#> [25]   6.3334420  17.0152696   1.0910408 -13.0665416   3.2040065  -7.4125398
#> [31]  -3.8109741   7.7526984 -16.7882805  15.5115488   5.4901616   2.7922357
#> [37]   5.5649430 -13.7856213  -7.2447039  -6.1999635   4.6681342  -1.3896594
#> [43]  -1.7993795   6.2200069  -8.7659656   1.8678588   2.4246791  -0.4903283
#> [49]  12.7134176   6.8712058

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4477792