Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -1.84110729 -1.16307135  1.14166742 -2.98218527  0.15237049 -2.35045845
#>  [7]  0.68252997 -1.12459619 -2.80149894 -0.44200265 -1.61805564  0.37191335
#> [13] -0.64833858 -3.55797123  0.55510200 -1.65762878 -0.72608487 -0.21000844
#> [19] -2.95297510 -1.74928566 -1.82830005  0.58603043  2.08730911 -1.99055439
#> [25] -1.34693316 -2.62521795 -1.23047720  0.91148226 -1.85548406 -2.21616174
#> [31] -2.51905690 -0.49115359  0.14081697 -2.09898009  0.51146829 -1.05340081
#> [37] -1.20810722 -1.98681802 -1.48537603  0.06878468 -3.23273093 -0.67324747
#> [43] -0.18214210 -1.04365459  0.02894215 -1.84636824  1.49923976  0.50314274
#> [49] -2.84726357  0.34125170

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 15.091925  8.987947  7.259024 13.846245 10.643978  8.970207 12.031482
#>  [8]  8.375200 12.378498 10.014555 10.543114  7.671816 11.604686 11.472006
#> [15]  7.102830 17.744901  8.754907  6.213156 14.086487 11.657148 11.356284
#> [22]  9.884468 10.031457 11.875416  5.952208 15.293932 12.480296 11.038644
#> [29] 10.354623  5.466235 14.258733  7.899309 14.227279 12.088169 13.576448
#> [36] 12.662580 17.235046  6.579391 10.264094 12.287950 17.282695  5.847617
#> [43]  8.159729 12.007796  7.252162 17.428936  8.991041 10.438376  4.931132
#> [50]  7.319626

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   1.383872461   9.529575331  -0.647016726 -18.652156974  14.670277668
#>  [6]  -7.377267475   0.139761875   0.649276619 -10.420700587  -7.526752917
#> [11]   0.356012287  -9.767347654  -8.620049318  -3.575486032   1.694348904
#> [16]  13.156022534  -0.005890924 -11.490528721   6.002485573   4.498291030
#> [21]   3.095826159 -12.466704573 -15.379684609  -2.586633664  -2.585919004
#> [26]  23.085657634   0.466328597  -7.494195031  -3.127177169  14.708201386
#> [31]  -0.513833874  10.620036064 -14.308622964  -0.525747047  -4.519206068
#> [36]  11.251389438   9.431122542 -12.648824085  -1.581808515  -4.582052680
#> [41]  -0.467800256  -1.879645704  -9.728802057  -6.396351839   0.364538026
#> [46]  -2.945762179  -5.907129370 -28.254512130   9.899329534  17.532405833

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4418737