Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -2.26517999  1.89423486  1.10419204 -1.84306370 -1.86572810  0.34874266
#>  [7]  1.32800880  1.32444761  2.38394060 -0.26481625 -0.20436792  1.40345532
#> [13] -0.09806728 -0.73877455 -2.35841659 -0.41172103  0.95587561  1.26827638
#> [19]  0.47535368 -0.50097535 -2.35534408  2.58068793  1.29373613  0.87121451
#> [25] -2.75702627 -1.90448334  2.20059259  0.77891303  2.04495600 -2.41918454
#> [31] -2.03673605  2.24252103 -0.60625485  1.22790824 -1.14257654 -2.00859743
#> [37]  0.44724145 -0.09286786 -1.73322930 -1.53047004 -0.67924810  0.40078268
#> [43]  0.55785695 -0.14949036 -1.23185842 -1.90341532  0.99217364 -0.22831074
#> [49]  2.21174962 -0.52156251

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  8.899532 10.269589 16.935654  5.993519  7.902830  8.465049  8.446118
#>  [8]  6.115814  8.058461  7.470240  5.589075  8.859759  4.565427  6.553889
#> [15]  5.391935  4.838744 15.124853  6.692197 11.802790  8.535483  9.149579
#> [22]  5.946759  7.757636  9.950262 12.374835  6.083952 10.293769  8.915959
#> [29]  7.813927  8.198603  5.761167  7.698627  3.798531  7.367082 10.662675
#> [36] 13.489658 10.439877  9.210918  9.360371  6.953311 12.170028 11.603225
#> [43]  8.274044 12.326737  6.084228  7.001233 11.916951  9.208485  6.618364
#> [50] 10.363097

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -6.0043815   0.2590955  -4.8615975   2.4586290  -2.0097461 -13.0391897
#>  [7]  -5.6707495  -5.0162359   0.2629586   1.1187745  -1.9619058   0.6277869
#> [13] -12.9800407  -5.3507869  -1.4275341  -3.2172645 -18.0631777   0.6364745
#> [19]  -4.4167509  -6.6934803  -4.2921786  -1.5520344   0.2305480   6.4471817
#> [25]   7.3399535   6.5009189  -2.9789604  19.3512347  -9.8030537  -1.4735811
#> [31] -11.6941060   1.5653703  -6.8868589  -4.8729290  -8.8629385  -5.1257345
#> [37]  -8.5070816  -7.7705989  -5.0559766 -13.5959610   8.5165942 -14.2422580
#> [43]   4.6073873   1.8859897  -3.4044077 -10.3313267   5.3269459  16.6706215
#> [49]  -8.2650777 -21.5593518

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4842105