Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.8335531 -2.1709217  0.7481985  0.5108628  2.7198904  1.1703137
#>  [7] -0.8342116  0.2234489  2.4403201  0.5978723 -0.3946634 -1.3442225
#> [13] -1.7699781  2.9298159  0.3861506  0.9933606 -1.2436514 -0.9370518
#> [19]  1.7335970  1.0612941 -1.0748862  0.0731955  0.2339531  0.7141313
#> [25]  1.4597996 -0.2413866  0.8933937 -0.7327083 -0.4471002  0.2816413
#> [31] -0.6815286 -0.1160264 -0.9091236  0.8049999  1.2490368 -1.5915402
#> [37] -0.3471948 -1.1062203  1.6231527  1.0674959 -1.7099906 -0.6720233
#> [43] -1.4907198  1.6312444  2.2415171  1.4718269 -2.0900936 -0.7364300
#> [49]  1.7745875  1.2626545

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  9.319820  8.332827 12.388596 13.570352  9.231456  6.385835 12.264658
#>  [8] 15.091763  6.806737 13.872763 13.778506 14.625829 10.623630 12.933266
#> [15] 11.546612  9.464070  8.138461 15.013423 12.844108 13.645809 13.530923
#> [22]  9.494038  9.983806  6.182625 10.012993 13.328491  9.301379 10.552742
#> [29]  9.585153 10.553999  8.502329 11.877623 12.307704  9.302819 15.421462
#> [36]  9.711772  5.430794 13.503200 10.683068  8.064859  9.634636 11.524349
#> [43]  9.340679  8.746763  5.280839  9.737427 12.163984 15.555772  9.059798
#> [50] 11.569421

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -1.8548968   6.9222134  -9.2159902 -14.4509718   7.8574068   4.0542213
#>  [7]   2.2438630   1.1020282   5.4080889  11.3610994   1.2987679   2.7213539
#> [13]  -5.5191849   1.0861604  28.6505154  -6.4983809  -9.9323395  -3.2469145
#> [19]  -8.1395261  16.9126977  -5.4816048   2.4666539  -9.4893856   0.3880862
#> [25]   6.8470859  -3.7527095  18.8791082  -3.0761071 -21.3516412  -8.6749208
#> [31]  -5.0937470  18.4737376  -0.4188688  -4.1652255   2.7780646   2.4956172
#> [37]   7.7121666   1.3229422  -3.0631687  21.6293914  -0.3091801  -5.3032035
#> [43]  -0.2109513  -4.4472904  18.3927345  -4.6456570  -5.7577482   0.9638412
#> [49]  -7.3289310  16.2041714

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.4321379