Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.49562288  0.09540458  1.11432388 -1.08524399  0.26957785 -1.17758192
#>  [7] -0.59331652  1.27179336 -0.08459929  1.32409222 -1.48156658  0.11101975
#> [13]  0.15216303 -0.04541650  1.08922965 -1.11994193  0.43056902  1.57140342
#> [19] -1.49416144  0.94966990 -0.07149111  0.37067579 -0.56678422 -0.83979230
#> [25]  1.22337483 -1.59482423  1.04307419  0.95366413 -0.72276764 -0.05246103
#> [31] -1.24250062 -0.02536418  0.43591445 -0.26190250  1.20278144 -1.52873883
#> [37]  0.62768825 -0.37048271 -0.76976915  0.64162729 -2.83981455  1.16330699
#> [43] -0.16022539 -0.68944846  2.13514281 -2.87861836 -1.10800282 -0.11510533
#> [49] -0.32733702  0.07384696

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 12.7046224 11.0094606 13.1143619 11.0120729  1.3336457 11.3300549
#>  [7]  9.8327880 13.3758583 17.7377900 -2.5119986  5.7496860 11.7346941
#> [13] 14.1966045 13.5740708 12.0856029 10.2566194  8.1446145 13.5682050
#> [19] 13.6555069  4.6660716 10.2935638 11.5823097 12.2484267 16.7337661
#> [25]  4.5277085 10.3383285  3.7256805 14.8174943 13.9884628  2.5162789
#> [31]  6.9133057  5.8172154 10.5305689 15.9977168  0.2810901  9.9221804
#> [37]  8.7228646 14.8757055 17.6860667  3.7701856  4.0645228 15.9932063
#> [43] 17.3150367 10.2646359  2.0782508 12.8405540 12.9071040 12.6126078
#> [49] 18.8862770  4.1432033

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   3.08313412 -21.80183227  -4.96225417   7.01820317  -0.84617566
#>  [6]  -0.25833207  -1.47849874   6.82075458   4.04594851   9.85562301
#> [11]  -0.55283362  -2.68359242 -21.05651046  -1.62165969   2.67631970
#> [16]  -4.67566581  -7.76934430  -1.50573773   4.59027055   5.41240354
#> [21]  -8.71868672 -15.59131436  -5.94530820   7.03396418   2.80113275
#> [26]  -9.37477153  -5.00520350   8.38462091  10.90291412  -0.41245580
#> [31]  -3.86089817  -0.04128571   6.40108502   7.56208148  15.35197785
#> [36]  -5.94057368  -9.41528107   0.51657743  -8.54448908  -8.99782554
#> [41]  -3.11407754 -10.52407954  -5.80202555  -3.49314974  15.77646880
#> [46]  -7.31968257 -18.14160352   9.88436934   2.73698220  -0.53172980

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3486835