Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.97336383 -0.21435724 -0.94085076  1.89513519 -1.50478637 -0.94310344
#>  [7] -0.53068553 -0.80205007  0.32667372 -0.31965020  1.05394991 -2.40983805
#> [13] -0.34345464 -1.23308293 -2.62355159 -1.40225115 -3.21200215 -0.40069434
#> [19]  1.52298721 -1.06151048 -0.21769845 -1.98522515 -0.69601834  2.16824419
#> [25]  1.24954034 -0.76153686 -2.06848832  1.34664780  0.94588192 -2.22528321
#> [31] -0.53427241 -0.10416666  0.24916327  2.21119630 -1.89550068 -1.17181924
#> [37] -1.90081919  0.90268236  2.24196782 -1.42315370  0.52439702 -0.64989583
#> [43] -0.21610608  1.54178599 -1.55738604 -1.66727407 -1.48664507 -0.06782305
#> [49]  2.35634173 -1.79026521

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  7.759311 22.586136 10.679327 12.859399  7.524312 12.163641 15.301796
#>  [8]  8.282420  9.935582 11.456785 12.036570  6.106705  4.435474 17.889886
#> [15]  6.167534  7.580276 14.763829 10.545669  9.679383 10.496021 10.068899
#> [22] 11.561206  6.819477 14.169339 10.043222 13.071947 14.161168  6.955511
#> [29] 12.185411  6.476890  3.717922  9.208036  9.225313 14.046844  8.580240
#> [36] 14.051879  6.731389  6.469711 14.067319 10.170340 11.363582 13.522567
#> [43]  6.839357 12.735646  6.926298  7.235705 12.067431  9.713468 12.454483
#> [50] 14.449652

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -0.3179462   1.0637048   2.1445615   4.3662482   5.4515077   4.2085255
#>  [7]  -8.9892431   8.6675997  15.5787856  -5.7491510   8.1435792  -0.8548041
#> [13] -16.8216005   9.7068005 -11.7363848   1.5629255 -19.8890987  10.6631481
#> [19]  17.4148236   3.3367654   1.1279478   9.2863578 -13.6125543  -3.2187751
#> [25]  -1.6798420  -1.9384192  -8.5777741  -9.1626819  18.6842380  10.1632474
#> [31]  -7.5005515  -5.4428851  -2.0641554  26.2870041  -2.9374059   3.8037540
#> [37] -11.2828373  -3.2019640  11.3445210   4.0065707  -4.2880162  -2.7813175
#> [43]  -7.1535729  12.4014159 -14.3978231  19.3311291 -17.3272793  -7.2136336
#> [49]  20.6443609  -9.4222410

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1101317