Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.49468402  0.72887071  2.28641085 -0.61271463 -1.98795658 -0.41644562
#>  [7] -0.71491148  2.15135632 -0.75594181 -0.79143911 -0.26776905 -1.05454880
#> [13]  2.05885802 -0.14256172 -0.64306085 -2.75261006  0.25419410  1.11077203
#> [19]  0.32945179  1.00941642  0.29057809  0.37432298  0.92453439  1.41284798
#> [25] -0.54926034 -0.33603791  0.17765989 -1.15707481  0.83457057  2.08144337
#> [31]  0.98526188  0.99713065  1.16962673 -0.60797791  0.02153174 -0.73455181
#> [37]  0.95823484  2.10922137 -0.81478341  0.29357149 -1.63356547 -0.45123070
#> [43]  1.62544391 -0.20543563 -1.09313754 -2.08928472 -1.33520805  2.46972613
#> [49]  0.06789094 -0.30229187

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.609606  4.816564 10.738095 12.364861 10.295222  8.095383  9.872944
#>  [8] 11.190487  6.460654  5.090839  5.514560 13.510429  6.382736  8.677252
#> [15]  7.738871  8.605687  6.610993 10.382482 14.435913 11.745273 11.665007
#> [22] 11.133306 11.522984 13.790320 13.169245  5.900879  8.205430 12.707683
#> [29] 10.831015 13.451675  5.801375 11.839111 13.160665  7.954571  8.105969
#> [36]  7.041726  5.330198 11.775901  9.630374  5.869594 10.373167  7.700830
#> [43]  9.398579  9.375292  6.718133 15.313971  5.888595 10.209103  9.801247
#> [50]  6.404085

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  17.87829157   8.72800184   0.05036082   0.92770677   3.85694053
#>  [6]   8.24797983 -11.96074323  -2.60403838   0.25587474  -5.32768906
#> [11]   3.36993308   4.44339224   6.05269951   1.34232374  -4.13718543
#> [16]   0.39429787   7.75737295   7.07987806  -5.23004457  -3.68772418
#> [21]   2.64860480  -4.38849814  -0.39460024 -10.94327448  13.46087925
#> [26]  15.95794733   0.14653378   7.04033012   1.45252451 -24.14350704
#> [31]  13.98972965 -15.50028964   0.04097300   1.60908003   7.17777440
#> [36]  -8.16398989   5.93844444   9.38467145   7.72953950  -8.12149323
#> [41]  -6.93112248  -0.46881918  11.87423189   4.11253149  -5.65840282
#> [46]   2.47437995  -4.79368340   1.82941998   3.50437355   0.72413950

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1895159