Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  2.66979209  3.01482361 -1.77012337  2.60854915  0.93849426 -0.56538625
#>  [7]  2.56313824  0.33622941 -0.08828176  1.56107129  3.60562128  2.81982359
#> [13] -0.84866065  2.14472858  0.81633973  2.75759533  3.26271847 -1.77281859
#> [19]  1.78614868 -0.79119226  1.88307511  4.04706560 -0.97003199  1.70782181
#> [25]  1.28238447  2.18936842  1.32466631 -1.34248432  1.38364191 -1.62125677
#> [31]  0.93347969  3.51668701 -0.23448296 -0.12111196 -1.36938933  0.69241090
#> [37]  2.16335833 -0.51301786  0.69424275  1.36450818  2.51051623  3.43288381
#> [43] -0.28160964 -0.84703061 -0.66243194  2.05477732  1.55873909 -2.54173655
#> [49]  0.65378372 -0.98631650

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.360845 15.125095  9.784611 11.274763 10.415013  5.833687  9.315940
#>  [8] 10.893640  8.392101  9.964072 -2.798015 11.999248  9.968284  9.686311
#> [15] 16.150149  5.821788  7.238853 14.157805  8.367835 10.682503  7.535734
#> [22] 10.192562  9.747194 10.022967 10.368811  8.887086 15.411632 13.369255
#> [29]  2.331600 11.076688 14.960130  9.880295  7.004696  9.005476 12.022476
#> [36]  4.564665  4.519541 11.228493 14.620752 13.264812  3.758942  8.393877
#> [43] 12.558727 10.572483 10.789418 14.905855  5.111034  8.084868  3.620698
#> [50] 10.809401

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -8.3655848  -4.2423812   5.3822127   6.5239968   9.8274899  -2.6801208
#>  [7]   9.2060780  -4.2181937  -9.9089208  -5.0179872   4.9842525  -6.0614804
#> [13]  -0.3539732  -8.0812078   3.6553544  -0.8032737 -19.2760947   1.5483700
#> [19]  -3.7351164  10.4458671 -10.8272971   1.9188107  -1.1098873   5.5729117
#> [25]  -9.0158101   9.2960280  -6.3818207   0.5853959  -8.1872211  -4.7917713
#> [31]   2.3840525   7.2592063  -2.5561016 -15.3503551   7.1970452  -0.2174185
#> [37]   3.4502520 -12.4889743  -3.3941001   5.3412160  -3.5215915 -11.1908442
#> [43]  -7.3016463   4.1115841   4.0797269  -0.6657116   1.2613336  -2.1038302
#> [49]  -0.5744656   6.0336342

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3572615