Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.44702154  1.92416987 -0.62609750 -1.11032539  0.14222465 -0.07628521
#>  [7] -1.17253246  0.92076736  0.54277957 -0.79209928 -0.85719297  0.81909978
#> [13]  1.67165362 -0.27929469 -2.22793493  0.09256576  0.26367893 -1.16097668
#> [19] -2.89504162  0.58700154  0.51764906  2.77175200 -0.62806557 -1.17503552
#> [25] -0.94657971  0.53153632 -0.84136211  2.23641498 -1.33362141 -1.71316990
#> [31] -1.11421409  1.86511211  0.89861120  1.26128441 -2.02671016  0.74160573
#> [37] -0.92726216  1.56434051  0.43364719 -1.81104429 -1.02829458 -1.11654281
#> [43]  1.88066753 -0.72850425 -1.63612995  1.19447387  0.19303951  0.89347938
#> [49]  0.34346675 -1.30234497

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.567227 16.051925 11.821800  9.516497  9.088042 10.462993 10.727842
#>  [8] 10.454819  7.467882 12.757579 12.551887 12.936596 11.786314  3.992377
#> [15] 16.522742  3.452619 12.692973 10.190605 13.271980  7.683118  8.858573
#> [22] 13.529824  7.977887  8.418725  5.400796  8.147773  6.929486 10.932379
#> [29] 12.859009 10.923959  6.505856 12.301146  5.386823  1.212449  8.622216
#> [36] 11.643631  8.827726 14.269304 10.797964  7.036960 10.206705 13.167040
#> [43] 12.034301  7.860937 17.213684 12.046374 16.205128  3.423930 10.344124
#> [50] 10.597175

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]   6.48945043  -8.00238173  -7.07297545  17.26041208  -2.92496045
#>  [6]  17.19416896  -3.09791226   3.12658585   1.59887750  -1.44758461
#> [11]  -0.89029575 -26.66708848  10.32561753 -14.02084199 -12.22096682
#> [16]  13.99385831  -9.22491608   2.33688990   4.35806967  -3.29220405
#> [21]  -0.86654768 -12.60858982   6.90400826   4.50765897   2.38048494
#> [26]  -6.14467291   8.59996007   0.75721588   5.22783218 -11.65399733
#> [31]  -3.49063406  14.01238362 -11.86642716  -4.62313490  16.36144562
#> [36]   9.25988820  -3.96954963   7.78724927   1.20509021  10.15511741
#> [41]  19.58554441  -9.48660387  -4.97413310  -0.08712577 -13.22512109
#> [46]  12.52750955  -3.26043049  -5.84957256  -6.33762665 -21.52405841

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1645034