Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.393767400  0.517705050 -0.810987097  1.749362611 -0.796681603
#>  [6]  0.046043630  1.019457879 -0.337703175 -0.890402519 -1.527428949
#> [11] -0.409308045  1.056967845 -2.376901375  1.134201479 -0.911500009
#> [16] -1.079569433  1.669578937 -2.707299997  0.266935017 -2.149857991
#> [21] -0.149310390  0.874037055 -1.382131613  1.456103811  0.892426976
#> [26] -0.219273004 -0.026618224 -1.302158916  0.066866430  0.499477861
#> [31] -0.009487967  2.711403105  0.917990082  0.707814019 -0.888141114
#> [36]  1.204481366  0.353231874 -1.844692708 -0.003159275  0.184553081
#> [41] -0.912503220  1.663746174  0.323333432  0.632528260 -0.771140686
#> [46]  1.818385312  0.365180100  0.041629523 -0.635549660  0.192560588

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 11.351470  6.151408 12.810490  6.649322 13.667573  8.238569  7.791227
#>  [8] 11.010738 12.980448 11.238872 14.824157 14.047812 14.481147  7.785106
#> [15] 10.961961 10.495331  8.904509 11.509073  8.525979  8.826737  9.044071
#> [22] 13.166059 18.743983 11.147588  8.046468 12.888073  6.134913 12.546603
#> [29]  9.014401 11.903998 15.050177  8.918048 13.029004 11.192107  8.187351
#> [36] 14.328897  8.389265 11.881117  8.858916 10.460602 11.325646 12.846646
#> [43] 14.517393  8.773055  7.342011 12.722398 17.927211  6.563157 11.948109
#> [50]  7.393556

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1] -24.4690298  -8.2928577  -4.6740897  -1.4616274 -12.8328108  -1.9881820
#>  [7]  -9.8280704  -1.9010750  12.6846250  11.8872317  -3.0479809 -15.8639102
#> [13]  11.0192421  10.2456360   0.3206113  -8.0080987  -6.1727048  -2.1141793
#> [19]  10.4392474  -4.5743806  -7.0607734 -20.9131901  -0.4716486   2.6297654
#> [25]  11.3595193 -16.4232052   0.3949697  -5.0834721  -0.6230639   7.5720629
#> [31]   1.6917168   2.6702332  11.8841328   9.7523988  -3.7997357  -4.6840988
#> [37]  -7.3994062   4.4745411   1.7507326  -7.5977312   9.9044306 -15.3059099
#> [43]   9.3298714   6.1526105  -4.9075000   1.9705324  -5.2466158  12.5969833
#> [49]   0.5199961   5.5765980

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.1756498