Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.18649018 -0.41022205 -1.54477242  0.94174660 -1.13806181 -1.02061875
#>  [7]  0.25698830 -1.33497122 -0.30299425 -1.42968433 -0.28062755  1.33618561
#> [13] -2.17741226  2.73503677  0.32098702  0.81401528  0.38171726 -0.35797195
#> [19] -0.03238229  0.75290110 -0.14387188  0.63016186 -0.06460025  1.42427844
#> [25] -0.06352103 -1.72053099  1.95715388 -0.23575706 -0.27448704 -1.05323825
#> [31] -0.09240763  2.02105671 -2.79259179  0.18952825  0.93659786 -1.45269920
#> [37] -0.04314192 -0.81822155  0.91987732 -3.17270470 -1.80593134  2.48826377
#> [43] -1.44157304  1.69800235  0.92356689 -1.73513898  3.45098501 -2.29344896
#> [49] -0.63896196  0.07754856

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  3.711139 13.626115 10.055193 15.931102 12.030887  7.739552 16.117180
#>  [8]  8.787141 13.573639  6.592492  7.669018  7.192118 14.330665 10.650668
#> [15]  4.264222 15.414665 10.163050  7.286853 13.920481 11.239665 11.707374
#> [22]  8.923459  8.648875 11.660296  8.704414  7.890451 12.980564  5.789933
#> [29] 14.299840  8.037610 12.110324  9.295600  6.270333 18.544723  9.088848
#> [36]  8.877822  9.244853  7.909755 10.263972 10.112774  5.496196 10.191012
#> [43] 11.078353 11.474131  4.430029  7.539777 13.489156  9.671225  8.692798
#> [50]  7.979540

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -3.2673575   6.7421868  -8.8211888   1.9587671   3.6064071 -13.5321878
#>  [7]   2.0168600   8.1754321  -6.3193741 -16.8088011  -7.9784961   1.6961912
#> [13]   5.7889575  -2.7146991   3.9523922  -6.0461352   7.2773509  -1.5418528
#> [19]   6.5074376  -0.1464291   3.6966261   0.4090239  -3.2400266   7.8389680
#> [25]   4.3368841  -3.4923331   5.5718183   9.5007947   1.1974534  -0.8150667
#> [31]  16.8671106  -4.1011130   3.0226223  -7.4942712   4.3599291  -2.1154574
#> [37]   6.7390718   2.2203364  11.8224744  -4.6442979  10.7690025  -6.9482157
#> [43]  -4.5871702   5.0875942  -5.1665452  -5.4956340  -4.6276531  16.0611601
#> [49] -10.1377676   9.5683443

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.01862833