Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1] -0.293323027  0.573271118  0.548582064  1.599718137 -0.164100988
#>  [6] -1.811364237 -0.092611093  1.769943758  1.012830815 -1.166676755
#> [11] -2.436937207  0.142036895  0.979481217  1.430821711  0.387477881
#> [16] -1.745401603  0.413079507  1.598434485 -2.377200457  0.773397375
#> [21]  0.341483158 -1.356355554 -0.095093305  0.774198974  2.045345340
#> [26] -2.890225092  0.803650745  0.672070205  0.905186634 -0.090011998
#> [31]  1.157839632 -1.978331940  0.838310314 -0.118452060  0.544533808
#> [36] -0.299162046  0.639811415 -0.902271797 -0.566208076  1.050558243
#> [41] -1.963809548  0.817355618  0.755011606  1.212866117 -1.152581167
#> [46] -0.009933684 -0.291592300  0.205637790  0.131338032 -1.188786470

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 13.595250 10.920241  8.528501 13.445219  5.913793  9.204603 11.497106
#>  [8] 10.452309 10.529992 17.863140  6.258114 15.230422  8.004381 14.949068
#> [15] 11.769698 10.592617  9.691386  7.440028  9.495727  9.239907  9.804295
#> [22]  6.298684 16.994291  9.086130 10.329725 13.806590 12.925234  9.318884
#> [29] 13.726794  8.014878  8.332474  7.029151 11.843443 11.741805  9.487179
#> [36] 12.954409 14.335209 12.322531  7.690812  9.076494  8.967633 10.517098
#> [43] 11.839003 12.672505  8.866003  9.181600 10.082654 13.705966 14.152345
#> [50]  8.821416

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -1.1364952  -0.2203239 -18.4763661   1.1811943  -0.8748809  -8.3934819
#>  [7]  -3.2853682 -18.9677993   1.1587951   4.3519362  -4.8516207  -2.8628755
#> [13]   0.7137850  -4.7988277  11.7704539  -7.9255103   4.8773663  -4.8080216
#> [19]  -4.0799905   5.2382641   1.6284004  -9.8030471  -2.2957022  -7.4478929
#> [25]   3.0494961   7.5306679 -12.8012395  -0.4351750 -15.4608305  -3.3538608
#> [31]  -6.4244851  -8.4766794  -2.2453327  -4.4918648  -7.4516752  -5.3035668
#> [37] -11.2999738  -7.8112465  -6.4013040  12.6276542  -0.2644465  -9.9719395
#> [43]  -7.6672053   9.7635285  -8.8105837   4.4208152 -13.0280766   2.9332435
#> [49]   6.8085316   8.9383533

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3815231