Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  1.22622913 -0.34294592  0.96094313  0.17294568 -0.43931599 -0.99610343
#>  [7]  0.47539756  0.79090931  0.21278149 -0.12040201 -0.82851577  0.22350421
#> [13]  1.46065834  0.47527934 -1.77697801 -1.00859367 -0.53841581  0.62348700
#> [19]  1.39425561 -0.08046841 -0.44893858 -0.72902370  0.87833417  2.42104823
#> [25] -0.04421837 -1.66302116 -2.59666555  1.67602738  0.41833720 -0.48897630
#> [31] -0.13148091  0.43435326  2.25876396  2.16568084 -1.80471115 -1.28263981
#> [37] -1.55436799  1.53548603  0.81689563 -0.38272269  0.17731128  1.53648035
#> [43]  0.25690320  1.07673186 -1.40121202 -3.14003117  0.52477285  1.77260148
#> [49]  1.92378725 -2.25397489

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1] 10.151068  8.253033 12.285479  9.200889 10.549166  8.153462  8.149379
#>  [8]  2.665322  6.166191  6.699887  8.542823  9.755117  8.192818  9.327492
#> [15] 10.808740  4.173444 17.002742  7.687218  8.485727 11.401026 13.985994
#> [22]  3.667657  8.965367  4.606096 12.426820  4.436790 10.050264  5.953594
#> [29]  6.706924 10.453946 13.996680 11.441564  8.971995  9.454622  9.953797
#> [36] 11.265545  6.420803 11.095330  9.367379 12.109268 13.533336  7.924989
#> [43]  8.894990  5.133789 12.504041  7.685673 14.016629  9.503507  4.268894
#> [50]  7.767893

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  13.2585655  18.5893890   7.6635521  12.4824994   8.6537276   9.8449488
#>  [7]  -2.1132937   4.3191377   4.1616707  -0.5781366  -3.7535009   2.3416230
#> [13]  17.6750780  15.5095400   4.3920840 -10.1539059 -10.1138445  -4.3201763
#> [19] -11.4685883  -4.9065298  -4.1394146  -9.9875998   8.1276302   8.5432112
#> [25]   1.8500304   8.3474527   0.9804517   0.7051742  -0.0831353   7.3813435
#> [31]  -4.8503478   0.6401301  -5.4981778   4.5020919  -9.8065101  -6.9664157
#> [37]   1.4847861  16.7097052   9.8463093   1.9618532   4.5784551   6.1549442
#> [43]   5.8615120   1.4497049   3.6641034  -7.1069046   3.4273359  17.2568825
#> [49]  22.6438594  11.5961462

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.3089112