Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

Usage

draw_normal_icc(
  mean = 0,
  N = NULL,
  clusters,
  sd = NULL,
  sd_between = NULL,
  total_sd = NULL,
  ICC = NULL
)

Arguments

mean

A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N

(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters

A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd

A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between

A number or vector of numbers, indicating the standard deviation between clusters.

total_sd

A number indicating the total sd of the resulting variable. May only be specified if ICC is specified and sd and sd_between are not.

ICC

A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)
#>  [1]  0.17903066  1.99128081 -0.07842873  0.47662356 -3.35274730 -0.34975603
#>  [7] -0.23021168 -0.41959771  1.52291313 -2.30546968 -0.29509711 -1.07605747
#> [13]  1.56191353 -0.15264980 -0.97530859 -0.04391100  0.48509900 -0.43372583
#> [19] -0.57821723 -0.70603948 -0.09855714  1.35690095  1.51993254  1.26261807
#> [25] -1.32598008  1.41698070  1.05848881 -1.17545808  0.90228585 -1.65777080
#> [31]  0.97817964  0.81670003  0.74660065 -0.02189591 -0.08530715 -1.83691069
#> [37] -1.78185590  0.50241885  0.10150851 -1.16615460 -1.17042222  1.83649482
#> [43]  0.02691967 -0.46020668 -1.35747900 -1.24248485 -0.80391712 -0.14467005
#> [49] -1.30388414 -2.05434561

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)
#>  [1]  6.951480 10.191989 10.142633 10.135666  5.360280  5.250864  7.079989
#>  [8] 15.014030 10.314530  6.789818  9.736932 13.084581 15.287867 13.914788
#> [15] 13.363141  8.817748 11.180213 14.596974  7.194184 13.299198  9.251778
#> [22] 12.636424 10.730809  9.120381  8.272959 12.463727 16.029261 10.937228
#> [29]  7.864302 10.864749 12.098784 12.659279 15.227028  6.366936 10.709288
#> [36]  7.111222 11.706431 11.308302 10.331858  9.097057  5.958805  9.564306
#> [43]  9.804135  7.065124  6.565818 12.848909  8.873639 14.489752  6.828761
#> [50]  7.849565

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)
#>  [1]  -5.2764576  -0.2667149   2.5899305 -25.9339148  -6.0184671  11.2662290
#>  [7]  -1.5873487  -3.5468973 -15.8329488  -6.8234657  10.4232334   6.4531973
#> [13]  -5.0709147   3.1850242   3.1607247  11.0340502  11.3914792   3.1043942
#> [19]  -3.5871186   3.0055914  -3.6783420  27.8353161   1.3617725   3.4147994
#> [25]   0.4982698   1.9538257  -4.8642652   1.2204471  -5.0989324  -6.8182040
#> [31]   3.1468910 -15.8962934  -2.9375080  -2.7130747 -10.4053466   4.9469606
#> [37]   1.0816308  -6.9086923   5.3052366 -16.1239993  -5.6313645   5.6716043
#> [43]  11.1679530 -10.1679681  -4.8662698   8.3938503   2.6343468   9.3525129
#> [49]  -1.0037378   4.1568670

# Can specify total SD instead:
total_sd_draw = draw_normal_icc(clusters = clusters, ICC = 0.5, total_sd = 3)
sd(total_sd_draw)
#> [1] 3

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared
#> [1] 0.107874