Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -0.442575979  1.812739008  1.198236838 -2.062864132 -0.941402685
#>  [6] -0.581228278 -0.500893195  1.878246612 -2.712600302 -1.761281744
#> [11] -0.562938612 -1.051168333  1.787571981 -1.084241615 -1.281496530
#> [16] -1.713941648  0.320496143 -0.569586640 -0.530921210 -1.453417137
#> [21]  0.037203708  0.366637014 -0.240792102 -1.983464088 -2.139392454
#> [26] -0.202276718 -0.846992857 -0.927704330 -2.023518590 -1.819803624
#> [31] -1.526971277  1.934447600  0.069561234 -3.023580463 -1.040382096
#> [36] -0.151988725  1.667056035  2.017855990 -1.829675167 -3.302810086
#> [41]  0.427191596 -1.439203115  0.003600543 -3.776046437 -1.585906244
#> [46] -2.397983852  0.240714245  1.287448219 -0.568112591 -0.470749417

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)

#>  [1]  5.5941073  3.8483030 11.1809905  7.5689486 10.8794589 -0.3795237
#>  [7]  4.3795722  9.1100963 11.1706698 10.4298950  6.1359899  5.2804414
#> [13] 10.8510957 12.8348254 13.4301442  8.4563239  4.3969631 10.0704476
#> [19] 12.8107250 10.2905950  5.0148561  7.4023980  8.4001266  8.1860124
#> [25] 10.9051650  5.4923991  3.8096746  9.4114961 15.6495637  7.3692963
#> [31] 10.1690230  5.1473703  6.4200299  8.1179109  2.2940903  6.4644038
#> [37]  6.6593270  9.0954370 17.3016105 14.7829235  8.0358768  2.8763618
#> [43]  8.0177048 13.2069961  4.2817715  5.5666968  4.1836925  4.5207289
#> [49]  9.0244261  9.9657273

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)

#>  [1]  -6.49360574 -14.53619562  -3.55181265 -19.12413980  17.53987273
#>  [6]  11.85791125  -0.04861590   5.79720862 -13.04238914  -5.77475915
#> [11]  -5.13901042   9.69454614   1.87342207 -16.52057714 -11.30955032
#> [16]  -1.41065956  -2.91610306  -1.50965287 -12.76442704 -19.68467893
#> [21]  -1.00747871  -0.06008108   3.56118727   8.17527872  -7.79805548
#> [26]   5.19609487  -3.63551174  -4.27589773  -8.34808155 -11.55143410
#> [31]   0.14245003   1.16396233  -4.99839597  -8.72474652  17.25490841
#> [36] -10.96654284  -5.48864939  -7.58877057  -9.02861913  -1.49360703
#> [41]   2.59201575  -8.19251532   1.58870309  -3.51660434 -15.84520168
#> [46]   8.61374519   8.30729371  -3.65113639  -9.10127581 -19.01496539

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared

#> [1] 0.4077254