Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation \(\rho\). The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(x = 0, N = NULL, clusters, sd = 1, rho = 0)

Arguments

x
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms

rho
A number indicating the desired ICC. If none is provided, will default to 0.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Examples

clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters)

#>  [1] -1.051 -0.946 -0.075  0.188  1.249 -0.549 -1.710 -0.701  1.898  1.114
#> [11]  0.847  0.391  1.054 -0.555 -0.629  0.303 -0.038  1.763  0.615 -1.661
#> [21] -1.048 -0.929  1.541  0.179 -1.345 -0.749 -0.989 -0.848 -0.910  0.410
#> [31] -2.820  0.325 -0.823 -0.091  0.494  1.042  0.094 -0.522 -0.510  0.578
#> [41] -2.067  0.531 -1.797 -0.324 -0.020  0.496 -0.316  1.669 -0.143  1.596