Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -1.02554589555  0.62443504160 -2.31939115286 -0.83443824172  0.55794551023
#>  [6]  1.53804682512 -0.22289234807  1.14690446594 -0.65298415590  2.17395191327
#> [11]  0.93017781910 -0.99045080885 -0.76061978747 -2.46877159639  1.19729849372
#> [16]  1.37574171129 -0.94850979858  0.51731006607 -1.17555276582  1.38167859134
#> [21] -0.47570412346 -0.74476058897 -0.12639826632 -0.26539863033  3.19550216981
#> [26]  0.30373914727  0.67777767630 -1.76081700260 -1.40004144012 -0.93182015081
#> [31]  0.45643319619 -1.46012657635 -1.09803959382  0.71703320070  1.03944986387
#> [36] -0.03334539499 -0.15605976028 -1.01039833808 -0.04367905912  0.80156012155
#> [41]  0.47585110047  0.22116711398 -0.15391440827  0.09899335053  1.49281216714
#> [46]  1.53738284473  0.47059084961 -1.67012691508  0.43175769886  1.57700317990

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)

#>  [1]  9.209538912  6.946955079  7.280875456  2.663388150 15.406216374
#>  [6] 13.109009661  9.931104635  6.439850524  7.340946553 11.535675521
#> [11] 11.219845289 10.889005790  9.943836409  6.589445159  7.329716280
#> [16] 11.209727037  8.794985299  8.253895225 10.731749801  7.722170255
#> [21]  9.411929542  6.296898979  9.374113345  8.337103771  6.355048089
#> [26] 10.574635252  7.147800959 10.207142467 12.951675524 12.653866245
#> [31] 10.357931418  8.776974347 10.735098329  7.420525160  9.172163731
#> [36] 15.434822568  4.686744939  8.902265146  8.857391358 12.050750197
#> [41]  6.447235044  9.464802213  4.916679553 14.715998765  7.217758120
#> [46] 11.054768770 12.617152323  5.155776608  4.699358851  7.700490168

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)

#>  [1]  13.6348820404  -9.3986100071 -10.1894901391  -4.3994492021  15.6367316440
#>  [6]   1.9925917184  -0.3076630385   7.6256721944   2.1741503757  17.2390779570
#> [11]   6.8659438811  -4.6678638151  -1.7383785970   5.5829176814   5.7807090978
#> [16]   5.6198046791 -12.8067743251  12.7308042072  -9.2940686647  12.7807675030
#> [21]   9.4914446968 -10.0633084774  10.5242160976   4.0939279532 -13.9777691778
#> [26]  -1.5699653734 -11.3426364373   9.7798558438  -1.1224243256  14.1309119449
#> [31]   7.6503184899 -13.0704663298   3.9990677369   6.2734685647   0.9106015725
#> [36]   7.0969074263 -17.4752527076  -6.4680374534  -0.3626732194  12.3058826153
#> [41]   2.3691357569 -19.1565832185   3.7634139973  -0.2841548969  -7.4997531889
#> [46]  -1.2743521664  -8.9896589122  -0.1177370598  -5.2378603460  -7.2251413074

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared

#> [1] 0.3853618356