Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(x = 0, N = NULL, clusters, sd = 1, ICC = 0)

Arguments

x
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms

ICC
A number indicating the desired ICC. If none is provided, will default to 0.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Examples

clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters)

#>  [1] -1.05057219 -0.94630671 -0.07517996  0.18792650  1.24910764 -0.54872690
#>  [7] -1.71033113 -0.70098598  1.89782704  1.11426860  0.84696175  0.39087260
#> [13]  1.05365534 -0.55548426 -0.62941268  0.30336454 -0.03803220  1.76272950
#> [19]  0.61462929 -1.66058679 -1.04845809 -0.92936450  1.54100663  0.17928638
#> [25] -1.34494038 -0.74890428 -0.98942495 -0.84796718 -0.91029604  0.40996180
#> [31] -2.82005204  0.32521078 -0.82326653 -0.09100570  0.49405516  1.04161559
#> [37]  0.09354927 -0.52194444 -0.50994176  0.57820522 -2.06716149  0.53088578
#> [43] -1.79744671 -0.32449648 -0.02025971  0.49643123 -0.31644161  1.66884891
#> [49] -0.14304240  1.59574669