Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -2.701267849 -1.262783164 -0.442575979  1.812739008  1.198236838
#>  [6] -1.470633724  1.642918693 -0.581228278 -0.500893195  1.878246612
#> [11] -2.120369894  0.823039634 -0.562938612 -1.051168333  1.787571981
#> [16] -0.492011207  1.302824848 -1.713941648  0.320496143 -0.569586640
#> [21]  0.061309198  1.130904241  0.037203708  0.366637014 -0.240792102
#> [26] -1.391233680  0.444928923 -0.202276718 -0.846992857 -0.927704330
#> [31] -1.431288182  0.764517753 -1.526971277  1.934447600  0.069561234
#> [36] -2.431350055  1.543939282 -0.151988725  1.667056035  2.017855990
#> [41] -1.237444759 -0.718488708  0.427191596 -1.439203115  0.003600543
#> [46] -3.183816029  0.998415134 -2.397983852  0.240714245  1.287448219

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)

#>  [1] 15.1921149 10.8756915  5.5941073  3.8483030 11.1809905  7.6751300
#>  [7] 15.0273218 -0.3795237  4.3795722  9.1100963 11.2768512 14.5777579
#> [13]  6.1359899  5.2804414 10.8510957 12.9410068 17.5780071  8.4563239
#> [19]  4.3969631 10.0704476 12.9169065 14.4384579  5.0148561  7.4023980
#> [25]  8.4001266  8.2921938 15.0530279  5.4923991  3.8096746  9.4114961
#> [31] 15.7557451 11.5171591 10.1690230  5.1473703  6.4200299  8.2240923
#> [37]  6.4419532  6.4644038  6.6593270  9.0954370 17.4077919 18.9307864
#> [43]  8.0358768  2.8763618  8.0177048 13.3131776  8.4296344  5.5666968
#> [49]  4.1836925  4.5207289

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)

#>  [1]  -9.73557085  -7.25203459  -6.49360574 -14.53619562  -3.55181265
#>  [6] -20.23193954  23.68657402  11.85791125  -0.04861590   5.79720862
#> [11] -14.15018889   0.37194214  -5.13901042   9.69454614   1.87342207
#> [16] -17.62837689  -5.16284904  -1.41065956  -2.91610306  -1.50965287
#> [21] -13.87222678 -13.53797765  -1.00747871  -0.06008108   3.56118727
#> [26]   7.06747897  -1.65135420   5.19609487  -3.63551174  -4.27589773
#> [31]  -9.45588129  -5.40473282   0.14245003   1.16396233  -4.99839597
#> [36]  -9.83254627  23.40160969 -10.96654284  -5.48864939  -7.58877057
#> [41] -10.13641887   4.65309425   2.59201575  -8.19251532   1.58870309
#> [46]  -4.62440408  -9.69850040   8.61374519   8.30729371  -3.65113639

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared

#> [1] 0.3250942