Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1]  1.325  0.606 -2.153  0.657  0.619  0.948  3.217 -3.402  1.442  0.108
#> [11] -0.112  0.686 -0.806  0.389  1.615  0.299 -0.907 -2.726 -1.029 -0.288
#> [21]  0.401  1.742  0.945  3.344 -0.555  0.584 -0.036 -2.008  0.295  1.147
#> [31]  0.305 -0.775 -1.806 -1.002  0.869  1.085  0.764 -0.329  0.645  1.071
#> [41]  0.530  0.215 -1.862  0.386  0.572  1.712  1.183  1.050  1.157 -1.534