Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation \(\rho\). The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(x = 0, N = NULL, cluster_ids, sd = 1, rho = 0.5)

Arguments

x
A number or vector of numbers, one mean per cluster.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(cluster_ids) if provided.

cluster_ids
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms

rho
A number indicating the desired RCC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Examples

cluster_ids = rep(1:5, 10)
draw_normal_icc(cluster_ids = cluster_ids)

#>  [1]  1.862 -1.927 -0.281  3.149  2.561  2.131 -3.112 -0.062  1.671 -1.166
#> [11]  0.824 -0.439 -0.512 -0.261 -0.616  1.506 -2.786 -1.564  0.712  1.644
#> [21]  1.105 -3.819 -0.996  2.119  2.657  0.792 -3.510 -1.747  2.332  2.515
#> [31]  0.808 -0.782 -0.158 -0.605  1.185  1.297 -1.097  0.665  1.782  0.869
#> [41]  1.984 -1.407 -1.038  2.188 -0.518  2.587 -3.262 -0.515  1.637  2.489