Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation \(\rho\). The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(x = 0, N = NULL, clusters, sd = 1, rho = 0)

Arguments

x
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms

rho
A number indicating the desired ICC. If none is provided, will default to 0.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Examples

clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters)

#>  [1] -0.78543266 -1.05673687 -0.79554143 -1.75627543 -0.69053790 -0.55854199
#>  [7] -0.53666333  0.22712713  0.97845492 -0.20888265 -1.39941046  0.25853729
#> [13] -0.44179945  0.56859986  2.12685046  0.42485844 -1.68428153  0.24940178
#> [19]  1.07283825  2.03936926  0.44945378  1.39181405  0.42656655  0.10758399
#> [25]  0.02229473  0.60361101 -0.26265057 -0.52826408  0.19214942 -1.14619967
#> [31]  0.84618466  0.08171963 -1.30511701 -0.94491206  0.45434159 -0.85520250
#> [37] -0.28689522  0.89496163  0.06730444 -0.16267634 -0.82731017  1.87650562
#> [43]  0.76644020  0.97995670  1.32178099 -1.11971083  0.51459982 -1.50909984
#> [49]  1.53274148  0.42914737