Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -1.599 -2.657 -0.776  2.086  2.363 -0.204 -0.555  0.978 -0.368  0.620
#> [11] -0.747 -0.984  1.688  0.803 -0.411 -2.099 -1.876  1.466  0.367 -0.096
#> [21] -1.799 -1.936 -0.923 -0.722  1.659 -3.871 -0.621 -0.898  0.097  1.743
#> [31] -0.009 -0.853 -0.597 -0.322  1.827 -3.118 -0.415 -1.873 -0.137  1.229
#> [41] -0.554 -1.263  1.594  0.045  2.845 -1.162 -2.030 -0.314 -1.771  1.868