Draw normal data with fixed intra-cluster correlation.

Data is generated to ensure inter-cluster correlation 0, intra-cluster correlation in expectation ICC. The data generating process used in this function is specified at the following URL: https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc

draw_normal_icc(mean = 0, N = NULL, clusters, sd = NULL,
  sd_between = NULL, ICC = NULL)

Arguments

mean
A number or vector of numbers, one mean per cluster. If none is provided, will default to 0.

N
(Optional) A number indicating the number of observations to be generated. Must be equal to length(clusters) if provided.

clusters
A vector of factors or items that can be coerced to clusters; the length will determine the length of the generated data.

sd
A number or vector of numbers, indicating the standard deviation of each cluster's error terms -- standard deviation within a cluster (default 1)

sd_between
A number or vector of numbers, indicating the standard deviation between clusters.

ICC
A number indicating the desired ICC.

Value

A vector of numbers corresponding to the observations from the supplied cluster IDs.

Details

The typical use for this function is for a user to provide an ICC and, optionally, a set of within-cluster standard deviations, sd. If the user does not provide sd, the default value is 1. These arguments imply a fixed between-cluster standard deviation.

An alternate mode for the function is to provide between-cluster standard deviations, sd_between, and an ICC. These arguments imply a fixed within-cluster standard deviation.

If users provide all three of ICC, sd_between, and sd, the function will warn the user and use the provided standard deviations for generating the data.

Examples

# Divide observations into clusters
clusters = rep(1:5, 10)

# Default: unit variance within each cluster
draw_normal_icc(clusters = clusters, ICC = 0.5)

#>  [1] -1.026  0.624 -2.319 -0.834  0.558  1.538 -0.223  1.147 -0.653  2.174
#> [11]  0.930 -0.990 -0.761 -2.469  1.197  1.376 -0.949  0.517 -1.176  1.382
#> [21] -0.476 -0.745 -0.126 -0.265  3.196  0.304  0.678 -1.761 -1.400 -0.932
#> [31]  0.456 -1.460 -1.098  0.717  1.039 -0.033 -0.156 -1.010 -0.044  0.802
#> [41]  0.476  0.221 -0.154  0.099  1.493  1.537  0.471 -1.670  0.432  1.577

# Alternatively, you can specify characteristics:
draw_normal_icc(mean = 10, clusters = clusters, sd = 3, ICC = 0.3)

#>  [1]  9.2  6.9  7.3  2.7 15.4 13.1  9.9  6.4  7.3 11.5 11.2 10.9  9.9  6.6  7.3
#> [16] 11.2  8.8  8.3 10.7  7.7  9.4  6.3  9.4  8.3  6.4 10.6  7.1 10.2 13.0 12.7
#> [31] 10.4  8.8 10.7  7.4  9.2 15.4  4.7  8.9  8.9 12.1  6.4  9.5  4.9 14.7  7.2
#> [46] 11.1 12.6  5.2  4.7  7.7

# Can specify between-cluster standard deviation instead:
draw_normal_icc(clusters = clusters, sd_between = 4, ICC = 0.2)

#>  [1]  13.63  -9.40 -10.19  -4.40  15.64   1.99  -0.31   7.63   2.17  17.24
#> [11]   6.87  -4.67  -1.74   5.58   5.78   5.62 -12.81  12.73  -9.29  12.78
#> [21]   9.49 -10.06  10.52   4.09 -13.98  -1.57 -11.34   9.78  -1.12  14.13
#> [31]   7.65 -13.07   4.00   6.27   0.91   7.10 -17.48  -6.47  -0.36  12.31
#> [41]   2.37 -19.16   3.76  -0.28  -7.50  -1.27  -8.99  -0.12  -5.24  -7.23

# Verify that ICC generated is accurate
corr_draw = draw_normal_icc(clusters = clusters, ICC = 0.4)
summary(lm(corr_draw ~ as.factor(clusters)))$r.squared

#> [1] 0.39