Fabricate data

fabricate helps you simulate a dataset before you collect it. You can either start with your own data and add simulated variables to it (by passing data to fabricate()) or start from scratch by defining N. Create hierarchical data with multiple levels of data such as citizens within cities within states using add_level() or modify existing hierarchical data using modify_level(). You can use any R function to create each variable. We provide several built-in options to easily draw from binary and count outcomes, including draw_binary, draw_count, draw_binary_icc, and draw_normal_icc.

fabricate(data = NULL, ..., N = NULL, ID_label = NULL)

add_level(N = NULL, ID_label = NULL, working_environment_ = NULL, ...,
  data_arguments = quos(...), nest = TRUE)

modify_level(N = NULL, ID_label = NULL, working_environment_ = NULL, ...,
  data_arguments = quos(...))

Arguments

data
(optional) user-provided data that forms the basis of the fabrication, i.e. you can add variables to existing data. Provide either N or data (N is the number of rows of the data if data is provided).

...
Variable or level-generating arguments, such as my_var = rnorm(N). For fabricate, you may also pass add_level() or modify_level() arguments, which define a level of a multi-level dataset. See examples.

N
(optional) number of units to draw. If provided as fabricate(N = 5), this determines the number of units in the single-level data. If provided in add_level, i.e. fabricate(cities = add_level(N = 5)), N determines the number of units in a specific level of a hierarchical dataset.

ID_label
(optional) variable name for ID variable, i.e. citizen_ID

working_environment_
Internal argument, not intended for end-user use.

data_arguments
Internal argument, not intended for end-user use.

nest
(Default TRUE) Boolean determining whether data in an add_level() call will be nested under the current working data frame or create a separate hierarchy of levels. See our vignette for cross-classified, non-nested data for details.

Value

data.frame

Examples

# Draw a single-level dataset with no covariates
df <- fabricate(N = 100)
head(df)

#>    ID
#> 1 001
#> 2 002
#> 3 003
#> 4 004
#> 5 005
#> 6 006

# Draw a single-level dataset with a covariate
df <- fabricate(
  N = 100,
  height_ft = runif(N, 3.5, 8)
)
head(df)

#>    ID height_ft
#> 1 001  5.550356
#> 2 002  4.490406
#> 3 003  4.126321
#> 4 004  4.472710
#> 5 005  5.325554
#> 6 006  4.551806

# Start with existing data
df <- fabricate(
  data = df,
  new_variable = rnorm(N)
)

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate(
 regions = add_level(N = 5),
 cities = add_level(N = 2, pollution = rnorm(N, mean = 5)))
head(df)

#>   regions cities pollution
#> 1       1     01  3.743827
#> 2       1     02  5.480848
#> 3       2     03  4.745469
#> 4       2     04  4.301791
#> 5       3     05  3.524879
#> 6       3     06  4.181729

# Start with existing data and add variables to hierarchical data
# at levels which are already present in the existing data.
# Note: do not provide N when adding variables to an existing level
df <- fabricate(
  data = df,
  regions = modify_level(watershed = sample(c(0, 1), N, replace = TRUE)),
  cities = modify_level(runoff = rnorm(N))
)

# For examples of cross-classified data, please read our vignette or check
# documentation for \code{cross_level}