Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x
Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions
A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep
Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.05041797  0.05041797  0.15041797
#> 2  02  1.20631082  1.20631082  1.30631082
#> 3  03  0.69514087  0.69514087  0.79514087
#> 4  04  0.39709527  0.39709527  0.49709527
#> 5  05  0.52502528  0.52502528  0.62502528
#> 6  06 -0.45748143 -0.45748143 -0.35748143
#> 7  07 -0.04519345 -0.04519345  0.05480655
#> 8  08 -0.77066745 -0.77066745 -0.67066745
#> 9  09 -1.02764037 -1.02764037 -0.92764037
#> 10 10  0.29310762  0.29310762  0.39310762

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)

#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.05065166 -0.05065166  0.04934834
#> 2  02 -0.73526176 -0.73526176 -0.63526176
#> 3  03  1.19918938  1.19918938  1.29918938
#> 4  04  0.82768475  0.82768475  0.92768475
#> 5  05 -0.26986130 -0.26986130 -0.16986130
#> 6  06 -2.02413727 -2.02413727 -1.92413727
#> 7  07  1.86554075  1.86554075  1.96554075
#> 8  08  0.58557803  0.58557803  0.68557803
#> 9  09 -0.98761646 -0.98761646 -0.88761646
#> 10 10  1.72909319  1.72909319  1.82909319

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)

#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.8575163  0.95751633  1.05751633  1.15751633
#> 2  02 -1.6643895 -1.56438947 -1.46438947 -1.36438947
#> 3  03  0.2109453  0.31094528  0.41094528  0.51094528
#> 4  04  0.8154844  0.91548439  1.01548439  1.11548439
#> 5  05  0.2166092  0.31660916  0.41660916  0.51660916
#> 6  06 -0.1629243 -0.06292434  0.03707566  0.13707566
#> 7  07  0.6307515  0.73075150  0.83075150  0.93075150
#> 8  08 -2.4443045 -2.34430451 -2.24430451 -2.14430451
#> 9  09 -0.3254595 -0.22545953 -0.12545953 -0.02545953
#> 10 10 -0.8068249 -0.70682489 -0.60682489 -0.50682489

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)

#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.3902926   1.3902926   1.4902926   1.6902926  2.29029262
#> 2  02 -0.5009814  -0.5009814  -0.4009814  -0.2009814  0.39901864
#> 3  03  0.9131139   0.9131139   1.0131139   1.2131139  1.81311389
#> 4  04  0.1770301   0.1770301   0.2770301   0.4770301  1.07703012
#> 5  05 -1.2027317  -1.2027317  -1.1027317  -0.9027317 -0.30273172
#> 6  06 -0.9962892  -0.9962892  -0.8962892  -0.6962892 -0.09628921
#> 7  07 -2.4418848  -2.4418848  -2.3418848  -2.1418848 -1.54188479
#> 8  08  0.3166672   0.3166672   0.4166672   0.6166672  1.21666723
#> 9  09  0.4130898   0.4130898   0.5130898   0.7130898  1.31308984
#> 10 10 -1.1579927  -1.1579927  -1.0579927  -0.8579927 -0.25799271