Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1547136 -0.1547136 -0.05471359
#> 2  02 -0.9556910 -0.9556910 -0.85569097
#> 3  03 -0.1918346 -0.1918346 -0.09183456
#> 4  04 -0.9226423 -0.9226423 -0.82264230
#> 5  05  0.4923921  0.4923921  0.59239209
#> 6  06  1.0789357  1.0789357  1.17893573
#> 7  07  0.3730910  0.3730910  0.47309100
#> 8  08  0.2958530  0.2958530  0.39585296
#> 9  09 -0.1613788 -0.1613788 -0.06137876
#> 10 10  2.7000988  2.7000988  2.80009878

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.44137114 -0.44137114 -0.34137114
#> 2  02  1.29054940  1.29054940  1.39054940
#> 3  03 -0.77334934 -0.77334934 -0.67334934
#> 4  04 -0.03486701 -0.03486701  0.06513299
#> 5  05 -0.43557439 -0.43557439 -0.33557439
#> 6  06 -0.30849430 -0.30849430 -0.20849430
#> 7  07  0.22917903  0.22917903  0.32917903
#> 8  08  0.13389101  0.13389101  0.23389101
#> 9  09 -0.75937594 -0.75937594 -0.65937594
#> 10 10 -0.64783255 -0.64783255 -0.54783255

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.25235659 -0.15235659 -0.05235659  0.04764341
#> 2  02  0.54551187  0.64551187  0.74551187  0.84551187
#> 3  03 -0.25276882 -0.15276882 -0.05276882  0.04723118
#> 4  04 -1.57584196 -1.47584196 -1.37584196 -1.27584196
#> 5  05 -0.03059724  0.06940276  0.16940276  0.26940276
#> 6  06 -0.49715523 -0.39715523 -0.29715523 -0.19715523
#> 7  07 -0.44076091 -0.34076091 -0.24076091 -0.14076091
#> 8  08  1.42953531  1.52953531  1.62953531  1.72953531
#> 9  09 -0.64930848 -0.54930848 -0.44930848 -0.34930848
#> 10 10 -0.37565237 -0.27565237 -0.17565237 -0.07565237

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.32987600 -0.32987600  -0.2298760 -0.02987600  0.57012400
#> 2  02 -0.36406119 -0.36406119  -0.2640612 -0.06406119  0.53593881
#> 3  03  1.79639907  1.79639907   1.8963991  2.09639907  2.69639907
#> 4  04  0.05212334  0.05212334   0.1521233  0.35212334  0.95212334
#> 5  05 -0.79675936 -0.79675936  -0.6967594 -0.49675936  0.10324064
#> 6  06  1.11313785  1.11313785   1.2131378  1.41313785  2.01313785
#> 7  07  0.27942512  0.27942512   0.3794251  0.57942512  1.17942512
#> 8  08 -0.65950235 -0.65950235  -0.5595024 -0.35950235  0.24049765
#> 9  09 -0.95649066 -0.95649066  -0.8564907 -0.65649066 -0.05649066
#> 10 10  0.16498767  0.16498767   0.2649877  0.46498767  1.06498767