Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.2552354  1.2552354  1.3552354
#> 2  02  0.6796211  0.6796211  0.7796211
#> 3  03  0.9776734  0.9776734  1.0776734
#> 4  04 -0.8728425 -0.8728425 -0.7728425
#> 5  05 -1.4745670 -1.4745670 -1.3745670
#> 6  06  2.8359358  2.8359358  2.9359358
#> 7  07  0.7610047  0.7610047  0.8610047
#> 8  08  0.5993678  0.5993678  0.6993678
#> 9  09 -0.8704454 -0.8704454 -0.7704454
#> 10 10 -0.7636942 -0.7636942 -0.6636942

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.17561150  0.17561150  0.2756115
#> 2  02 -1.00560508 -1.00560508 -0.9056051
#> 3  03 -0.02693980 -0.02693980  0.0730602
#> 4  04  1.04058321  1.04058321  1.1405832
#> 5  05  0.43737399  0.43737399  0.5373740
#> 6  06 -1.41182734 -1.41182734 -1.3118273
#> 7  07  0.45859933  0.45859933  0.5585993
#> 8  08  0.08446411  0.08446411  0.1844641
#> 9  09  0.35297121  0.35297121  0.4529712
#> 10 10  0.45381505  0.45381505  0.5538150

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -1.9048720 -1.80487203 -1.70487203 -1.60487203
#> 2  02 -1.0357884 -0.93578838 -0.83578838 -0.73578838
#> 3  03  1.3594571  1.45945710  1.55945710  1.65945710
#> 4  04 -0.1875019 -0.08750186  0.01249814  0.11249814
#> 5  05 -0.3907806 -0.29078056 -0.19078056 -0.09078056
#> 6  06 -0.3128200 -0.21281997 -0.11281997 -0.01281997
#> 7  07 -2.3693969 -2.26939695 -2.16939695 -2.06939695
#> 8  08 -0.1891522 -0.08915220  0.01084780  0.11084780
#> 9  09  0.5994174  0.69941736  0.79941736  0.89941736
#> 10 10  0.2867814  0.38678144  0.48678144  0.58678144

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.09213496  0.09213496   0.1921350   0.3921350   0.9921350
#> 2  02 -1.41758644 -1.41758644  -1.3175864  -1.1175864  -0.5175864
#> 3  03  0.45059105  0.45059105   0.5505911   0.7505911   1.3505911
#> 4  04  0.70493054  0.70493054   0.8049305   1.0049305   1.6049305
#> 5  05 -0.42906895 -0.42906895  -0.3290690  -0.1290690   0.4709310
#> 6  06  0.66022223  0.66022223   0.7602222   0.9602222   1.5602222
#> 7  07 -1.03916734 -1.03916734  -0.9391673  -0.7391673  -0.1391673
#> 8  08  0.13003987  0.13003987   0.2300399   0.4300399   1.0300399
#> 9  09 -1.79395198 -1.79395198  -1.6939520  -1.4939520  -0.8939520
#> 10 10  0.11420613  0.11420613   0.2142061   0.4142061   1.0142061