Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.3953564 -1.3953564 -1.2953564
#> 2  02 -1.3670680 -1.3670680 -1.2670680
#> 3  03  0.2102745  0.2102745  0.3102745
#> 4  04 -0.0834838 -0.0834838  0.0165162
#> 5  05  1.7511354  1.7511354  1.8511354
#> 6  06  0.6525982  0.6525982  0.7525982
#> 7  07  0.3553846  0.3553846  0.4553846
#> 8  08  0.8955796  0.8955796  0.9955796
#> 9  09  0.7338238  0.7338238  0.8338238
#> 10 10 -0.6088465 -0.6088465 -0.5088465

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.7987059 -0.7987059 -0.6987059
#> 2  02  0.8661398  0.8661398  0.9661398
#> 3  03  0.3937013  0.3937013  0.4937013
#> 4  04 -1.3378407 -1.3378407 -1.2378407
#> 5  05 -0.5494015 -0.5494015 -0.4494015
#> 6  06  1.3343723  1.3343723  1.4343723
#> 7  07  0.9395192  0.9395192  1.0395192
#> 8  08  0.1232514  0.1232514  0.2232514
#> 9  09 -0.7592652 -0.7592652 -0.6592652
#> 10 10 -3.0569322 -3.0569322 -2.9569322

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.24043693 -0.14043693 -0.04043693  0.05956307
#> 2  02 -1.48698231 -1.38698231 -1.28698231 -1.18698231
#> 3  03 -0.27389681 -0.17389681 -0.07389681  0.02610319
#> 4  04 -0.04788123  0.05211877  0.15211877  0.25211877
#> 5  05 -0.87942666 -0.77942666 -0.67942666 -0.57942666
#> 6  06  0.35860696  0.45860696  0.55860696  0.65860696
#> 7  07  0.56784616  0.66784616  0.76784616  0.86784616
#> 8  08 -0.01343555  0.08656445  0.18656445  0.28656445
#> 9  09 -0.85207264 -0.75207264 -0.65207264 -0.55207264
#> 10 10 -0.62235940 -0.52235940 -0.42235940 -0.32235940

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.62649329 -0.62649329  -0.5264933  -0.3264933   0.2735067
#> 2  02  0.26198176  0.26198176   0.3619818   0.5619818   1.1619818
#> 3  03 -1.20975270 -1.20975270  -1.1097527  -0.9097527  -0.3097527
#> 4  04 -0.75999145 -0.75999145  -0.6599915  -0.4599915   0.1400085
#> 5  05 -1.60126845 -1.60126845  -1.5012684  -1.3012684  -0.7012684
#> 6  06  0.78880579  0.78880579   0.8888058   1.0888058   1.6888058
#> 7  07  0.66651434  0.66651434   0.7665143   0.9665143   1.5665143
#> 8  08  1.03907122  1.03907122   1.1390712   1.3390712   1.9390712
#> 9  09  0.56757282  0.56757282   0.6675728   0.8675728   1.4675728
#> 10 10  0.02842842  0.02842842   0.1284284   0.3284284   0.9284284