Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8253717 -0.8253717 -0.7253717
#> 2  02 -0.9539752 -0.9539752 -0.8539752
#> 3  03  0.5123489  0.5123489  0.6123489
#> 4  04  1.0570720  1.0570720  1.1570720
#> 5  05  0.3467438  0.3467438  0.4467438
#> 6  06  0.3640067  0.3640067  0.4640067
#> 7  07  1.3504305  1.3504305  1.4504305
#> 8  08  0.1367972  0.1367972  0.2367972
#> 9  09 -0.2353528 -0.2353528 -0.1353528
#> 10 10 -0.6924870 -0.6924870 -0.5924870

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.55239871  0.55239871  0.65239871
#> 2  02 -0.24817956 -0.24817956 -0.14817956
#> 3  03 -0.08432593 -0.08432593  0.01567407
#> 4  04 -1.00866456 -1.00866456 -0.90866456
#> 5  05  1.24773343  1.24773343  1.34773343
#> 6  06  0.79293162  0.79293162  0.89293162
#> 7  07 -2.02721015 -2.02721015 -1.92721015
#> 8  08 -0.21002474 -0.21002474 -0.11002474
#> 9  09 -0.05351323 -0.05351323  0.04648677
#> 10 10  0.41576417  0.41576417  0.51576417

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.8499164  0.9499164  1.0499164  1.1499164
#> 2  02 -1.3999923 -1.2999923 -1.1999923 -1.0999923
#> 3  03 -1.1944659 -1.0944659 -0.9944659 -0.8944659
#> 4  04 -0.9953398 -0.8953398 -0.7953398 -0.6953398
#> 5  05  0.5920577  0.6920577  0.7920577  0.8920577
#> 6  06 -1.5132504 -1.4132504 -1.3132504 -1.2132504
#> 7  07  0.3796585  0.4796585  0.5796585  0.6796585
#> 8  08  1.3648234  1.4648234  1.5648234  1.6648234
#> 9  09 -0.4227735 -0.3227735 -0.2227735 -0.1227735
#> 10 10  0.2357829  0.3357829  0.4357829  0.5357829

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.2859409   1.2859409  1.385940941   1.5859409   2.1859409
#> 2  02 -0.0949094  -0.0949094  0.005090599   0.2050906   0.8050906
#> 3  03  0.8167146   0.8167146  0.916714621   1.1167146   1.7167146
#> 4  04  1.3011166   1.3011166  1.401116578   1.6011166   2.2011166
#> 5  05 -1.4766159  -1.4766159 -1.376615865  -1.1766159  -0.5766159
#> 6  06  0.5136203   0.5136203  0.613620338   0.8136203   1.4136203
#> 7  07 -1.2797445  -1.2797445 -1.179744503  -0.9797445  -0.3797445
#> 8  08 -1.1557800  -1.1557800 -1.055780022  -0.8557800  -0.2557800
#> 9  09  1.7478228   1.7478228  1.847822793   2.0478228   2.6478228
#> 10 10  0.9796807   0.9796807  1.079680720   1.2796807   1.8796807