Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.64167467  0.64167467  0.74167467
#> 2  02  0.36469878  0.36469878  0.46469878
#> 3  03  0.90578267  0.90578267  1.00578267
#> 4  04  0.56995937  0.56995937  0.66995937
#> 5  05 -0.06236584 -0.06236584  0.03763416
#> 6  06 -1.79210734 -1.79210734 -1.69210734
#> 7  07 -0.20722601 -0.20722601 -0.10722601
#> 8  08 -0.24851600 -0.24851600 -0.14851600
#> 9  09  0.52156973  0.52156973  0.62156973
#> 10 10  0.56490625  0.56490625  0.66490625

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.3752679 -1.3752679 -1.2752679
#> 2  02 -0.5099980 -0.5099980 -0.4099980
#> 3  03 -0.7604340 -0.7604340 -0.6604340
#> 4  04 -2.2150523 -2.2150523 -2.1150523
#> 5  05  0.7194486  0.7194486  0.8194486
#> 6  06 -1.1480524 -1.1480524 -1.0480524
#> 7  07 -1.1076129 -1.1076129 -1.0076129
#> 8  08  1.4716893  1.4716893  1.5716893
#> 9  09  0.9723131  0.9723131  1.0723131
#> 10 10 -0.6767941 -0.6767941 -0.5767941

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1        Y_Z_2       Y_Z_3
#> 1  01  1.83524402  1.93524402  2.035244023  2.13524402
#> 2  02 -1.52916707 -1.42916707 -1.329167072 -1.22916707
#> 3  03 -0.11350956 -0.01350956  0.086490436  0.18649044
#> 4  04  1.64996476  1.74996476  1.849964759  1.94996476
#> 5  05  2.64640878  2.74640878  2.846408781  2.94640878
#> 6  06  0.09008709  0.19008709  0.290087094  0.39008709
#> 7  07 -0.90399351 -0.80399351 -0.703993514 -0.60399351
#> 8  08 -0.20396923 -0.10396923 -0.003969229  0.09603077
#> 9  09  1.53899294  1.63899294  1.738992942  1.83899294
#> 10 10 -0.34026787 -0.24026787 -0.140267871 -0.04026787

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.3336037  -0.3336037  -0.2336037  -0.0336037   0.5663963
#> 2  02  0.6508903   0.6508903   0.7508903   0.9508903   1.5508903
#> 3  03  0.8045160   0.8045160   0.9045160   1.1045160   1.7045160
#> 4  04 -0.4999626  -0.4999626  -0.3999626  -0.1999626   0.4000374
#> 5  05 -0.9174706  -0.9174706  -0.8174706  -0.6174706  -0.0174706
#> 6  06  0.3619591   0.3619591   0.4619591   0.6619591   1.2619591
#> 7  07 -1.5493282  -1.5493282  -1.4493282  -1.2493282  -0.6493282
#> 8  08  0.7760669   0.7760669   0.8760669   1.0760669   1.6760669
#> 9  09 -1.0937982  -1.0937982  -0.9937982  -0.7937982  -0.1937982
#> 10 10  3.2874219   3.2874219   3.3874219   3.5874219   4.1874219