Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.08893996  1.08893996  1.1889400
#> 2  02 -0.43702681 -0.43702681 -0.3370268
#> 3  03  0.31023750  0.31023750  0.4102375
#> 4  04  0.69331936  0.69331936  0.7933194
#> 5  05  1.10303728  1.10303728  1.2030373
#> 6  06 -0.34275330 -0.34275330 -0.2427533
#> 7  07  0.45558749  0.45558749  0.5555875
#> 8  08  0.05123967  0.05123967  0.1512397
#> 9  09  0.90857765  0.90857765  1.0085777
#> 10 10 -1.64559840 -1.64559840 -1.5455984

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.6964435 -0.6964435 -0.59644351
#> 2  02 -0.3900455 -0.3900455 -0.29004549
#> 3  03 -0.4896042 -0.4896042 -0.38960425
#> 4  04  1.8977752  1.8977752  1.99777518
#> 5  05 -0.6817055 -0.6817055 -0.58170552
#> 6  06  0.4793176  0.4793176  0.57931757
#> 7  07 -0.1612664 -0.1612664 -0.06126637
#> 8  08  1.6358581  1.6358581  1.73585814
#> 9  09  0.7528165  0.7528165  0.85281648
#> 10 10 -0.3892436 -0.3892436 -0.28924355

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U        Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  0.8845244  0.984524413  1.08452441  1.1845244
#> 2  02 -2.6616185 -2.561618544 -2.46161854 -2.3616185
#> 3  03 -1.2479453 -1.147945335 -1.04794533 -0.9479453
#> 4  04  0.3726093  0.472609251  0.57260925  0.6726093
#> 5  05 -0.5756199 -0.475619875 -0.37561988 -0.2756199
#> 6  06 -0.6257410 -0.525741030 -0.42574103 -0.3257410
#> 7  07 -1.8111183 -1.711118265 -1.61111826 -1.5111183
#> 8  08 -1.7050668 -1.605066780 -1.50506678 -1.4050668
#> 9  09 -0.1098127 -0.009812653  0.09018735  0.1901873
#> 10 10  2.3592647  2.459264658  2.55926466  2.6592647

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.1262807   0.1262807   0.2262807   0.4262807   1.0262807
#> 2  02 -0.7505846  -0.7505846  -0.6505846  -0.4505846   0.1494154
#> 3  03  1.1648317   1.1648317   1.2648317   1.4648317   2.0648317
#> 4  04 -1.0882300  -1.0882300  -0.9882300  -0.7882300  -0.1882300
#> 5  05 -0.4982167  -0.4982167  -0.3982167  -0.1982167   0.4017833
#> 6  06  0.8999439   0.8999439   0.9999439   1.1999439   1.7999439
#> 7  07  2.3002053   2.3002053   2.4002053   2.6002053   3.2002053
#> 8  08  1.0461322   1.0461322   1.1461322   1.3461322   1.9461322
#> 9  09  1.2342168   1.2342168   1.3342168   1.5342168   2.1342168
#> 10 10  0.6392016   0.6392016   0.7392016   0.9392016   1.5392016