Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01  0.078153335  0.078153335  0.1781533
#> 2  02 -0.446130878 -0.446130878 -0.3461309
#> 3  03 -1.558479480 -1.558479480 -1.4584795
#> 4  04  0.752513039  0.752513039  0.8525130
#> 5  05 -1.312335537 -1.312335537 -1.2123355
#> 6  06  0.873472487  0.873472487  0.9734725
#> 7  07 -1.024520677 -1.024520677 -0.9245207
#> 8  08  1.677349286  1.677349286  1.7773493
#> 9  09 -1.818455409 -1.818455409 -1.7184554
#> 10 10  0.004123124  0.004123124  0.1041231

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01 -0.551137444 -0.551137444 -0.45113744
#> 2  02  0.459243574  0.459243574  0.55924357
#> 3  03 -0.006118999 -0.006118999  0.09388100
#> 4  04 -0.071173221 -0.071173221  0.02882678
#> 5  05  0.441913254  0.441913254  0.54191325
#> 6  06 -0.592827280 -0.592827280 -0.49282728
#> 7  07 -0.115632832 -0.115632832 -0.01563283
#> 8  08 -0.567090442 -0.567090442 -0.46709044
#> 9  09 -0.598944819 -0.598944819 -0.49894482
#> 10 10 -2.075681321 -2.075681321 -1.97568132

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID             U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.1343555228  0.23435552  0.33435552  0.43435552
#> 2  02 -0.1287373269 -0.02873733  0.07126267  0.17126267
#> 3  03  0.6159419425  0.71594194  0.81594194  0.91594194
#> 4  04  1.1337314117  1.23373141  1.33373141  1.43373141
#> 5  05  0.0004559876  0.10045599  0.20045599  0.30045599
#> 6  06  1.2737614308  1.37376143  1.47376143  1.57376143
#> 7  07 -2.5767724604 -2.47677246 -2.37677246 -2.27677246
#> 8  08 -0.3378560322 -0.23785603 -0.13785603 -0.03785603
#> 9  09  0.0487476392  0.14874764  0.24874764  0.34874764
#> 10 10  0.1352337980  0.23523380  0.33523380  0.43523380

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.7881909   0.7881909   0.8881909   1.0881909   1.6881909
#> 2  02 -0.6825367  -0.6825367  -0.5825367  -0.3825367   0.2174633
#> 3  03  0.3274812   0.3274812   0.4274812   0.6274812   1.2274812
#> 4  04 -1.5503610  -1.5503610  -1.4503610  -1.2503610  -0.6503610
#> 5  05  0.3537159   0.3537159   0.4537159   0.6537159   1.2537159
#> 6  06 -0.6421712  -0.6421712  -0.5421712  -0.3421712   0.2578288
#> 7  07 -1.1344981  -1.1344981  -1.0344981  -0.8344981  -0.2344981
#> 8  08  1.2979497   1.2979497   1.3979497   1.5979497   2.1979497
#> 9  09  0.5451108   0.5451108   0.6451108   0.8451108   1.4451108
#> 10 10 -1.1480319  -1.1480319  -1.0480319  -0.8480319  -0.2480319