Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0         Y_Z_1
#> 1  01  0.83800151  0.83800151  0.9380015068
#> 2  02 -0.69648051 -0.69648051 -0.5964805129
#> 3  03 -0.06644050 -0.06644050  0.0335595032
#> 4  04 -0.62559339 -0.62559339 -0.5255933887
#> 5  05  0.27966649  0.27966649  0.3796664945
#> 6  06  0.33534259  0.33534259  0.4353425922
#> 7  07  0.33370118  0.33370118  0.4337011765
#> 8  08  1.77399986  1.77399986  1.8739998562
#> 9  09 -0.09952709 -0.09952709  0.0004729085
#> 10 10 -0.35701022 -0.35701022 -0.2570102199

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.26440424  0.26440424  0.36440424
#> 2  02  1.10139483  1.10139483  1.20139483
#> 3  03  0.38266381  0.38266381  0.48266381
#> 4  04  0.54704257  0.54704257  0.64704257
#> 5  05 -0.59631354 -0.59631354 -0.49631354
#> 6  06  0.06554535  0.06554535  0.16554535
#> 7  07 -0.02271585 -0.02271585  0.07728415
#> 8  08  0.54875289  0.54875289  0.64875289
#> 9  09 -0.98974146 -0.98974146 -0.88974146
#> 10 10 -1.28983513 -1.28983513 -1.18983513

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01 -0.5921738 -0.4921738 -0.39217376 -0.292173760
#> 2  02 -0.2978642 -0.1978642 -0.09786421  0.002135785
#> 3  03  0.8313146  0.9313146  1.03131464  1.131314644
#> 4  04 -1.5155098 -1.4155098 -1.31550981 -1.215509814
#> 5  05 -0.5531959 -0.4531959 -0.35319588 -0.253195876
#> 6  06  1.4271246  1.5271246  1.62712461  1.727124610
#> 7  07  0.2126995  0.3126995  0.41269947  0.512699470
#> 8  08  0.8322599  0.9322599  1.03225987  1.132259870
#> 9  09 -1.0148413 -0.9148413 -0.81484126 -0.714841259
#> 10 10 -0.5394413 -0.4394413 -0.33944134 -0.239441337

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.3226968  -0.3226968  -0.2226968 -0.02269677  0.57730323
#> 2  02 -0.9908264  -0.9908264  -0.8908264 -0.69082636 -0.09082636
#> 3  03  1.2124781   1.2124781   1.3124781  1.51247815  2.11247815
#> 4  04  0.4187741   0.4187741   0.5187741  0.71877405  1.31877405
#> 5  05 -2.9256615  -2.9256615  -2.8256615 -2.62566148 -2.02566148
#> 6  06 -0.5252490  -0.5252490  -0.4252490 -0.22524904  0.37475096
#> 7  07  2.0115000   2.0115000   2.1115000  2.31150000  2.91150000
#> 8  08  0.5171481   0.5171481   0.6171481  0.81714809  1.41714809
#> 9  09  1.1454259   1.1454259   1.2454259  1.44542592  2.04542592
#> 10 10 -0.9558411  -0.9558411  -0.8558411 -0.65584107 -0.05584107