Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.53824155  1.53824155  1.6382416
#> 2  02 -0.28092326 -0.28092326 -0.1809233
#> 3  03  0.03329421  0.03329421  0.1332942
#> 4  04  1.27777932  1.27777932  1.3777793
#> 5  05 -1.26780307 -1.26780307 -1.1678031
#> 6  06  1.16983217  1.16983217  1.2698322
#> 7  07 -0.35817270 -0.35817270 -0.2581727
#> 8  08 -2.46370300 -2.46370300 -2.3637030
#> 9  09 -0.93833079 -0.93833079 -0.8383308
#> 10 10 -0.75232754 -0.75232754 -0.6523275

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.10730009 -1.10730009 -1.00730009
#> 2  02 -1.15674861 -1.15674861 -1.05674861
#> 3  03  0.10327849  0.10327849  0.20327849
#> 4  04 -0.71290486 -0.71290486 -0.61290486
#> 5  05 -1.26793502 -1.26793502 -1.16793502
#> 6  06 -0.03557386 -0.03557386  0.06442614
#> 7  07 -0.08727435 -0.08727435  0.01272565
#> 8  08  0.02819314  0.02819314  0.12819314
#> 9  09  0.23745005  0.23745005  0.33745005
#> 10 10 -0.30652687 -0.30652687 -0.20652687

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.33971319  0.43971319  0.5397132  0.63971319
#> 2  02  0.54688787  0.64688787  0.7468879  0.84688787
#> 3  03 -0.51883906 -0.41883906 -0.3188391 -0.21883906
#> 4  04  2.25689872  2.35689872  2.4568987  2.55689872
#> 5  05 -0.67084782 -0.57084782 -0.4708478 -0.37084782
#> 6  06  0.80249749  0.90249749  1.0024975  1.10249749
#> 7  07 -0.31465139 -0.21465139 -0.1146514 -0.01465139
#> 8  08  0.21114774  0.31114774  0.4111477  0.51114774
#> 9  09  1.24589557  1.34589557  1.4458956  1.54589557
#> 10 10 -0.05542595  0.04457405  0.1445740  0.24457405

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.4109022  -0.4109022 -0.31090221 -0.11090221   0.4890978
#> 2  02 -0.1169315  -0.1169315 -0.01693145  0.18306855   0.7830685
#> 3  03  0.0294703   0.0294703  0.12947030  0.32947030   0.9294703
#> 4  04  0.9067940   0.9067940  1.00679404  1.20679404   1.8067940
#> 5  05 -3.2596626  -3.2596626 -3.15966260 -2.95966260  -2.3596626
#> 6  06 -0.2094895  -0.2094895 -0.10948953  0.09051047   0.6905105
#> 7  07 -0.4515243  -0.4515243 -0.35152434 -0.15152434   0.4484757
#> 8  08 -0.5207380  -0.5207380 -0.42073797 -0.22073797   0.3792620
#> 9  09 -0.3686963  -0.3686963 -0.26869625 -0.06869625   0.5313037
#> 10 10 -1.5312045  -1.5312045 -1.43120445 -1.23120445  -0.6312045