Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8066709 -0.8066709 -0.7066709
#> 2  02 -1.1080601 -1.1080601 -1.0080601
#> 3  03  0.1330966  0.1330966  0.2330966
#> 4  04  0.7446013  0.7446013  0.8446013
#> 5  05  1.6188413  1.6188413  1.7188413
#> 6  06  0.7911886  0.7911886  0.8911886
#> 7  07  1.1864356  1.1864356  1.2864356
#> 8  08 -0.9357956 -0.9357956 -0.8357956
#> 9  09 -0.2838839 -0.2838839 -0.1838839
#> 10 10 -0.6121508 -0.6121508 -0.5121508

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  2.13500552  2.13500552  2.23500552
#> 2  02  0.03336281  0.03336281  0.13336281
#> 3  03  1.12049144  1.12049144  1.22049144
#> 4  04 -0.05692692 -0.05692692  0.04307308
#> 5  05  0.75001086  0.75001086  0.85001086
#> 6  06 -1.43847324 -1.43847324 -1.33847324
#> 7  07  0.93420758  0.93420758  1.03420758
#> 8  08 -0.71277285 -0.71277285 -0.61277285
#> 9  09  0.52561460  0.52561460  0.62561460
#> 10 10 -0.63451839 -0.63451839 -0.53451839

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.3694708  0.4694708  0.5694708  0.66947082
#> 2  02  0.3214868  0.4214868  0.5214868  0.62148682
#> 3  03  1.7079726  1.8079726  1.9079726  2.00797258
#> 4  04  0.5611025  0.6611025  0.7611025  0.86110251
#> 5  05 -0.3332750 -0.2332750 -0.1332750 -0.03327498
#> 6  06 -0.5331101 -0.4331101 -0.3331101 -0.23311012
#> 7  07  0.5111103  0.6111103  0.7111103  0.81111034
#> 8  08 -1.5050640 -1.4050640 -1.3050640 -1.20506401
#> 9  09  0.7270314  0.8270314  0.9270314  1.02703136
#> 10 10 -1.2055693 -1.1055693 -1.0055693 -0.90556930

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.490351436 -0.490351436 -0.39035144  -0.1903514   0.4096486
#> 2  02  0.683241759  0.683241759  0.78324176   0.9832418   1.5832418
#> 3  03 -1.803564835 -1.803564835 -1.70356484  -1.5035648  -0.9035648
#> 4  04  0.261831236  0.261831236  0.36183124   0.5618312   1.1618312
#> 5  05 -1.443217129 -1.443217129 -1.34321713  -1.1432171  -0.5432171
#> 6  06  0.008825049  0.008825049  0.10882505   0.3088250   0.9088250
#> 7  07 -1.220969583 -1.220969583 -1.12096958  -0.9209696  -0.3209696
#> 8  08  0.801199056  0.801199056  0.90119906   1.1011991   1.7011991
#> 9  09 -0.197341280 -0.197341280 -0.09734128   0.1026587   0.7026587
#> 10 10 -0.454166077 -0.454166077 -0.35416608  -0.1541661   0.4458339