Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.0402137  1.0402137  1.1402137
#> 2  02  0.9155304  0.9155304  1.0155304
#> 3  03  0.1403775  0.1403775  0.2403775
#> 4  04 -0.9899433 -0.9899433 -0.8899433
#> 5  05 -1.2226165 -1.2226165 -1.1226165
#> 6  06  1.6472742  1.6472742  1.7472742
#> 7  07  2.5593298  2.5593298  2.6593298
#> 8  08 -1.4467240 -1.4467240 -1.3467240
#> 9  09 -0.3927248 -0.3927248 -0.2927248
#> 10 10  1.7408661  1.7408661  1.8408661

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1130461 -1.1130461 -1.0130461
#> 2  02  0.2253152  0.2253152  0.3253152
#> 3  03  0.7297758  0.7297758  0.8297758
#> 4  04  0.1637518  0.1637518  0.2637518
#> 5  05  0.4332010  0.4332010  0.5332010
#> 6  06 -1.5395363 -1.5395363 -1.4395363
#> 7  07  0.5465717  0.5465717  0.6465717
#> 8  08 -0.6930818 -0.6930818 -0.5930818
#> 9  09 -0.5343415 -0.5343415 -0.4343415
#> 10 10  0.7198485  0.7198485  0.8198485

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.38322549  0.4832255  0.5832255  0.68322549
#> 2  02 -0.70555161 -0.6055516 -0.5055516 -0.40555161
#> 3  03  0.46905288  0.5690529  0.6690529  0.76905288
#> 4  04  0.60340258  0.7034026  0.8034026  0.90340258
#> 5  05  1.61027910  1.7102791  1.8102791  1.91027910
#> 6  06 -0.31653481 -0.2165348 -0.1165348 -0.01653481
#> 7  07  0.52510147  0.6251015  0.7251015  0.82510147
#> 8  08  1.29359278  1.3935928  1.4935928  1.59359278
#> 9  09  0.06942744  0.1694274  0.2694274  0.36942744
#> 10 10  1.19264991  1.2926499  1.3926499  1.49264991

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.20609213 -0.20609213  -0.1060921  0.09390787   0.6939079
#> 2  02 -0.55354265 -0.55354265  -0.4535426 -0.25354265   0.3464574
#> 3  03  1.09815195  1.09815195   1.1981520  1.39815195   1.9981520
#> 4  04  2.46557805  2.46557805   2.5655780  2.76557805   3.3655780
#> 5  05 -0.74160023 -0.74160023  -0.6416002 -0.44160023   0.1583998
#> 6  06  0.49503803  0.49503803   0.5950380  0.79503803   1.3950380
#> 7  07  0.24258822  0.24258822   0.3425882  0.54258822   1.1425882
#> 8  08  0.18705472  0.18705472   0.2870547  0.48705472   1.0870547
#> 9  09  0.06027042  0.06027042   0.1602704  0.36027042   0.9602704
#> 10 10  1.09666951  1.09666951   1.1966695  1.39666951   1.9966695