Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.5338410 -1.5338410 -1.4338410
#> 2  02 -0.4430632 -0.4430632 -0.3430632
#> 3  03  1.2338736  1.2338736  1.3338736
#> 4  04 -0.5190966 -0.5190966 -0.4190966
#> 5  05  0.0552759  0.0552759  0.1552759
#> 6  06 -0.5495589 -0.5495589 -0.4495589
#> 7  07 -0.5723131 -0.5723131 -0.4723131
#> 8  08  0.1474922  0.1474922  0.2474922
#> 9  09 -0.7443007 -0.7443007 -0.6443007
#> 10 10 -0.5650549 -0.5650549 -0.4650549

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.721367425  0.721367425  0.82136742
#> 2  02 -1.184252877 -1.184252877 -1.08425288
#> 3  03 -0.004659127 -0.004659127  0.09534087
#> 4  04  0.305306901  0.305306901  0.40530690
#> 5  05 -0.372498924 -0.372498924 -0.27249892
#> 6  06 -0.539890462 -0.539890462 -0.43989046
#> 7  07  1.124589513  1.124589513  1.22458951
#> 8  08  1.291760249  1.291760249  1.39176025
#> 9  09 -0.909261337 -0.909261337 -0.80926134
#> 10 10  1.622786450  1.622786450  1.72278645

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.5197507 -0.4197507 -0.31975067 -0.21975067
#> 2  02 -0.5746016 -0.4746016 -0.37460162 -0.27460162
#> 3  03 -1.3653465 -1.2653465 -1.16534647 -1.06534647
#> 4  04 -0.7492677 -0.6492677 -0.54926774 -0.44926774
#> 5  05 -0.4275733 -0.3275733 -0.22757325 -0.12757325
#> 6  06 -0.2664876 -0.1664876 -0.06648757  0.03351243
#> 7  07 -0.3320163 -0.2320163 -0.13201626 -0.03201626
#> 8  08 -1.7890640 -1.6890640 -1.58906402 -1.48906402
#> 9  09  0.9453154  1.0453154  1.14531541  1.24531541
#> 10 10  0.9798180  1.0798180  1.17981800  1.27981800

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.76193143 -0.76193143 -0.66193143  -0.4619314   0.1380686
#> 2  02 -0.63763257 -0.63763257 -0.53763257  -0.3376326   0.2623674
#> 3  03 -0.03775627 -0.03775627  0.06224373   0.2622437   0.8622437
#> 4  04  0.51429090  0.51429090  0.61429090   0.8142909   1.4142909
#> 5  05 -0.73436441 -0.73436441 -0.63436441  -0.4343644   0.1656356
#> 6  06 -1.44371912 -1.44371912 -1.34371912  -1.1437191  -0.5437191
#> 7  07  1.55246856  1.55246856  1.65246856   1.8524686   2.4524686
#> 8  08 -2.19171930 -2.19171930 -2.09171930  -1.8917193  -1.2917193
#> 9  09  1.66772058  1.66772058  1.76772058   1.9677206   2.5677206
#> 10 10 -2.39370799 -2.39370799 -2.29370799  -2.0937080  -1.4937080