Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.1917496  0.1917496  0.2917496
#> 2  02 -0.6670923 -0.6670923 -0.5670923
#> 3  03 -1.5343998 -1.5343998 -1.4343998
#> 4  04 -0.8028032 -0.8028032 -0.7028032
#> 5  05 -0.4824648 -0.4824648 -0.3824648
#> 6  06 -0.8879462 -0.8879462 -0.7879462
#> 7  07 -0.7433881 -0.7433881 -0.6433881
#> 8  08  0.7086233  0.7086233  0.8086233
#> 9  09 -0.5794070 -0.5794070 -0.4794070
#> 10 10  2.6201544  2.6201544  2.7201544

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.91593671 -0.91593671 -0.8159367
#> 2  02  1.05750175  1.05750175  1.1575017
#> 3  03  1.10385171  1.10385171  1.2038517
#> 4  04  0.05085496  0.05085496  0.1508550
#> 5  05 -1.64346211 -1.64346211 -1.5434621
#> 6  06  0.75182010  0.75182010  0.8518201
#> 7  07  0.76520949  0.76520949  0.8652095
#> 8  08  0.61587579  0.61587579  0.7158758
#> 9  09 -0.28468542 -0.28468542 -0.1846854
#> 10 10 -0.68981225 -0.68981225 -0.5898122

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.3473704 -0.24737043 -0.14737043 -0.04737043
#> 2  02 -0.2650006 -0.16500055 -0.06500055  0.03499945
#> 3  03  1.9350371  2.03503708  2.13503708  2.23503708
#> 4  04 -0.6864510 -0.58645099 -0.48645099 -0.38645099
#> 5  05  2.4342090  2.53420896  2.63420896  2.73420896
#> 6  06 -0.2596850 -0.15968501 -0.05968501  0.04031499
#> 7  07 -0.6857135 -0.58571354 -0.48571354 -0.38571354
#> 8  08 -1.0033316 -0.90333164 -0.80333164 -0.70333164
#> 9  09 -0.1468687 -0.04686867  0.05313133  0.15313133
#> 10 10 -0.6830046 -0.58300462 -0.48300462 -0.38300462

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.509357415  1.509357415  1.60935741  1.80935741  2.40935741
#> 2  02 -0.364423925 -0.364423925 -0.26442392 -0.06442392  0.53557608
#> 3  03  2.253543499  2.253543499  2.35354350  2.55354350  3.15354350
#> 4  04 -0.005803251 -0.005803251  0.09419675  0.29419675  0.89419675
#> 5  05  1.493480112  1.493480112  1.59348011  1.79348011  2.39348011
#> 6  06  2.089747637  2.089747637  2.18974764  2.38974764  2.98974764
#> 7  07 -0.809071380 -0.809071380 -0.70907138 -0.50907138  0.09092862
#> 8  08  0.285006398  0.285006398  0.38500640  0.58500640  1.18500640
#> 9  09  1.624848761  1.624848761  1.72484876  1.92484876  2.52484876
#> 10 10 -0.127658795 -0.127658795 -0.02765879  0.17234121  0.77234121