Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.4467079  0.4467079  0.5467079
#> 2  02  0.4223444  0.4223444  0.5223444
#> 3  03  0.2301386  0.2301386  0.3301386
#> 4  04  0.1276919  0.1276919  0.2276919
#> 5  05  0.7138850  0.7138850  0.8138850
#> 6  06  1.0006763  1.0006763  1.1006763
#> 7  07 -0.6348639 -0.6348639 -0.5348639
#> 8  08  0.2944840  0.2944840  0.3944840
#> 9  09 -0.4894692 -0.4894692 -0.3894692
#> 10 10 -1.2805502 -1.2805502 -1.1805502

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.7388030  1.7388030  1.83880297
#> 2  02  0.8455389  0.8455389  0.94553895
#> 3  03  0.9280530  0.9280530  1.02805296
#> 4  04 -0.4708018 -0.4708018 -0.37080176
#> 5  05  0.8560131  0.8560131  0.95601314
#> 6  06  0.9129884  0.9129884  1.01298837
#> 7  07  1.1296742  1.1296742  1.22967419
#> 8  08 -0.1136123 -0.1136123 -0.01361234
#> 9  09 -1.8099483 -1.8099483 -1.70994832
#> 10 10 -0.3419150 -0.3419150 -0.24191496

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1         Y_Z_2       Y_Z_3
#> 1  01  1.2985779  1.3985779  1.4985778584  1.59857786
#> 2  02 -1.3581729 -1.2581729 -1.1581729151 -1.05817292
#> 3  03 -1.0801698 -0.9801698 -0.8801698380 -0.78016984
#> 4  04  0.6013811  0.7013811  0.8013811158  0.90138112
#> 5  05  0.6113877  0.7113877  0.8113877316  0.91138773
#> 6  06  0.7045364  0.8045364  0.9045363614  1.00453636
#> 7  07 -0.8968465 -0.7968465 -0.6968464702 -0.59684647
#> 8  08 -0.2001835 -0.1001835 -0.0001834537  0.09981655
#> 9  09 -1.0047501 -0.9047501 -0.8047501243 -0.70475012
#> 10 10  1.2654295  1.3654295  1.4654295420  1.56542954

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.6046534  -0.6046534  -0.5046534 -0.30465343   0.2953466
#> 2  02 -1.5284553  -1.5284553  -1.4284553 -1.22845532  -0.6284553
#> 3  03 -1.3231313  -1.3231313  -1.2231313 -1.02313133  -0.4231313
#> 4  04 -1.9674005  -1.9674005  -1.8674005 -1.66740046  -1.0674005
#> 5  05 -1.1586117  -1.1586117  -1.0586117 -0.85861175  -0.2586117
#> 6  06  0.2156543   0.2156543   0.3156543  0.51565428   1.1156543
#> 7  07 -0.5011718  -0.5011718  -0.4011718 -0.20117176   0.3988282
#> 8  08 -0.3956719  -0.3956719  -0.2956719 -0.09567191   0.5043281
#> 9  09  1.4061963   1.4061963   1.5061963  1.70619630   2.3061963
#> 10 10 -0.5379949  -0.5379949  -0.4379949 -0.23799494   0.3620051