Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.9850309 -0.9850309 -0.8850309
#> 2  02 -0.3473567 -0.3473567 -0.2473567
#> 3  03 -0.7660973 -0.7660973 -0.6660973
#> 4  04 -0.3762333 -0.3762333 -0.2762333
#> 5  05  1.5011809  1.5011809  1.6011809
#> 6  06  0.5440986  0.5440986  0.6440986
#> 7  07  0.4587301  0.4587301  0.5587301
#> 8  08 -0.0827839 -0.0827839  0.0172161
#> 9  09  1.7215857  1.7215857  1.8215857
#> 10 10  1.8642645  1.8642645  1.9642645

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.5082561  0.5082561  0.6082561
#> 2  02 -1.8026496 -1.8026496 -1.7026496
#> 3  03 -1.2018014 -1.2018014 -1.1018014
#> 4  04 -0.3673915 -0.3673915 -0.2673915
#> 5  05 -0.7862963 -0.7862963 -0.6862963
#> 6  06 -1.4354375 -1.4354375 -1.3354375
#> 7  07 -1.3119939 -1.3119939 -1.2119939
#> 8  08 -0.8332900 -0.8332900 -0.7332900
#> 9  09  1.9061379  1.9061379  2.0061379
#> 10 10  0.2017093  0.2017093  0.3017093

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -2.230245798 -2.1302458 -2.0302458 -1.9302458
#> 2  02 -0.847827626 -0.7478276 -0.6478276 -0.5478276
#> 3  03  0.008445913  0.1084459  0.2084459  0.3084459
#> 4  04  1.591888736  1.6918887  1.7918887  1.8918887
#> 5  05  1.925785885  2.0257859  2.1257859  2.2257859
#> 6  06 -0.662190165 -0.5621902 -0.4621902 -0.3621902
#> 7  07 -0.762305277 -0.6623053 -0.5623053 -0.4623053
#> 8  08 -0.562097739 -0.4620977 -0.3620977 -0.2620977
#> 9  09 -0.645036131 -0.5450361 -0.4450361 -0.3450361
#> 10 10 -0.696424342 -0.5964243 -0.4964243 -0.3964243

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.1722321   0.1722321   0.2722321   0.4722321  1.07223210
#> 2  02 -0.8675058  -0.8675058  -0.7675058  -0.5675058  0.03249421
#> 3  03  1.6265704   1.6265704   1.7265704   1.9265704  2.52657037
#> 4  04 -1.0980975  -1.0980975  -0.9980975  -0.7980975 -0.19809752
#> 5  05 -0.6895204  -0.6895204  -0.5895204  -0.3895204  0.21047957
#> 6  06 -1.1141831  -1.1141831  -1.0141831  -0.8141831 -0.21418309
#> 7  07 -0.2509041  -0.2509041  -0.1509041   0.0490959  0.64909590
#> 8  08 -2.0455269  -2.0455269  -1.9455269  -1.7455269 -1.14552688
#> 9  09 -0.8883587  -0.8883587  -0.7883587  -0.5883587  0.01164126
#> 10 10  0.4174818   0.4174818   0.5174818   0.7174818  1.31748179