Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01 -1.12027912 -1.12027912 -1.020279119
#> 2  02 -0.32326717 -0.32326717 -0.223267169
#> 3  03  0.13046696  0.13046696  0.230466960
#> 4  04  1.09768831  1.09768831  1.197688309
#> 5  05 -1.33750083 -1.33750083 -1.237500827
#> 6  06  0.63993094  0.63993094  0.739930937
#> 7  07  0.03883083  0.03883083  0.138830828
#> 8  08  2.50706781  2.50706781  2.607067810
#> 9  09  0.14317826  0.14317826  0.243178258
#> 10 10 -0.10842029 -0.10842029 -0.008420288

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.2256029 -1.2256029 -1.1256029
#> 2  02  0.2794678  0.2794678  0.3794678
#> 3  03  0.9741182  0.9741182  1.0741182
#> 4  04  0.6581341  0.6581341  0.7581341
#> 5  05 -0.3824126 -0.3824126 -0.2824126
#> 6  06  0.2914944  0.2914944  0.3914944
#> 7  07  0.7174869  0.7174869  0.8174869
#> 8  08 -0.5684218 -0.5684218 -0.4684218
#> 9  09 -0.7262683 -0.7262683 -0.6262683
#> 10 10  2.7803540  2.7803540  2.8803540

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.52983291  0.6298329  0.7298329  0.8298329
#> 2  02 -1.04856868 -0.9485687 -0.8485687 -0.7485687
#> 3  03 -1.61543837 -1.5154384 -1.4154384 -1.3154384
#> 4  04 -2.69153823 -2.5915382 -2.4915382 -2.3915382
#> 5  05 -0.49654720 -0.3965472 -0.2965472 -0.1965472
#> 6  06  0.11199627  0.2119963  0.3119963  0.4119963
#> 7  07 -0.95205020 -0.8520502 -0.7520502 -0.6520502
#> 8  08  1.11349903  1.2134990  1.3134990  1.4134990
#> 9  09  0.09021733  0.1902173  0.2902173  0.3902173
#> 10 10  0.13191272  0.2319127  0.3319127  0.4319127

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.8395153  -1.8395153  -1.7395153  -1.5395153  -0.9395153
#> 2  02 -0.5771025  -0.5771025  -0.4771025  -0.2771025   0.3228975
#> 3  03  0.3965998   0.3965998   0.4965998   0.6965998   1.2965998
#> 4  04 -1.5223505  -1.5223505  -1.4223505  -1.2223505  -0.6223505
#> 5  05 -0.4259406  -0.4259406  -0.3259406  -0.1259406   0.4740594
#> 6  06 -1.1094705  -1.1094705  -1.0094705  -0.8094705  -0.2094705
#> 7  07 -0.5277101  -0.5277101  -0.4277101  -0.2277101   0.3722899
#> 8  08 -0.4115759  -0.4115759  -0.3115759  -0.1115759   0.4884241
#> 9  09  0.1434244   0.1434244   0.2434244   0.4434244   1.0434244
#> 10 10  0.7910290   0.7910290   0.8910290   1.0910290   1.6910290