Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.99893204 -0.99893204 -0.89893204
#> 2  02  0.53890165  0.53890165  0.63890165
#> 3  03 -1.10755941 -1.10755941 -1.00755941
#> 4  04  1.86605539  1.86605539  1.96605539
#> 5  05  0.40638928  0.40638928  0.50638928
#> 6  06 -0.05926883 -0.05926883  0.04073117
#> 7  07 -0.33192394 -0.33192394 -0.23192394
#> 8  08 -0.46110734 -0.46110734 -0.36110734
#> 9  09 -2.23662393 -2.23662393 -2.13662393
#> 10 10  0.01162736  0.01162736  0.11162736

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.35375942  0.35375942  0.45375942
#> 2  02  0.90213419  0.90213419  1.00213419
#> 3  03  0.75843105  0.75843105  0.85843105
#> 4  04 -1.34126434 -1.34126434 -1.24126434
#> 5  05  0.05867754  0.05867754  0.15867754
#> 6  06  0.53992160  0.53992160  0.63992160
#> 7  07  1.79160934  1.79160934  1.89160934
#> 8  08  0.74028011  0.74028011  0.84028011
#> 9  09  0.80346076  0.80346076  0.90346076
#> 10 10 -0.11473946 -0.11473946 -0.01473946

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01  1.94321501  2.04321501  2.14321501  2.243215007
#> 2  02 -0.30587032 -0.20587032 -0.10587032 -0.005870316
#> 3  03 -0.13328333 -0.03328333  0.06671667  0.166716667
#> 4  04 -0.26814347 -0.16814347 -0.06814347  0.031856534
#> 5  05  0.84563187  0.94563187  1.04563187  1.145631867
#> 6  06 -1.45488712 -1.35488712 -1.25488712 -1.154887124
#> 7  07 -0.01154022  0.08845978  0.18845978  0.288459777
#> 8  08  0.77853797  0.87853797  0.97853797  1.078537974
#> 9  09  0.81478984  0.91478984  1.01478984  1.114789844
#> 10 10  0.88090216  0.98090216  1.08090216  1.180902162

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.33065753 -1.33065753  -1.2306575 -1.03065753  -0.4306575
#> 2  02  0.57732389  0.57732389   0.6773239  0.87732389   1.4773239
#> 3  03 -0.35961723 -0.35961723  -0.2596172 -0.05961723   0.5403828
#> 4  04  0.08487534  0.08487534   0.1848753  0.38487534   0.9848753
#> 5  05 -1.05246909 -1.05246909  -0.9524691 -0.75246909  -0.1524691
#> 6  06  0.92674782  0.92674782   1.0267478  1.22674782   1.8267478
#> 7  07 -0.52332397 -0.52332397  -0.4233240 -0.22332397   0.3766760
#> 8  08  1.13617348  1.13617348   1.2361735  1.43617348   2.0361735
#> 9  09 -0.71178242 -0.71178242  -0.6117824 -0.41178242   0.1882176
#> 10 10  0.13361522  0.13361522   0.2336152  0.43361522   1.0336152