Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.49202973  1.49202973  1.59202973
#> 2  02  0.30349850  0.30349850  0.40349850
#> 3  03  1.68626914  1.68626914  1.78626914
#> 4  04 -1.20944830 -1.20944830 -1.10944830
#> 5  05  0.46671523  0.46671523  0.56671523
#> 6  06  0.22096237  0.22096237  0.32096237
#> 7  07  1.05462445  1.05462445  1.15462445
#> 8  08  0.26334416  0.26334416  0.36334416
#> 9  09 -1.25631415 -1.25631415 -1.15631415
#> 10 10 -0.06247364 -0.06247364  0.03752636

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.05207534 -1.05207534 -0.95207534
#> 2  02 -2.20815445 -2.20815445 -2.10815445
#> 3  03  0.61636685  0.61636685  0.71636685
#> 4  04 -0.63829874 -0.63829874 -0.53829874
#> 5  05  1.40625669  1.40625669  1.50625669
#> 6  06 -0.38228374 -0.38228374 -0.28228374
#> 7  07 -0.23061625 -0.23061625 -0.13061625
#> 8  08 -0.51575452 -0.51575452 -0.41575452
#> 9  09 -0.07871476 -0.07871476  0.02128524
#> 10 10 -1.67957410 -1.67957410 -1.57957410

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.57003160 -0.4700316 -0.3700316 -0.2700316
#> 2  02 -1.25209533 -1.1520953 -1.0520953 -0.9520953
#> 3  03  0.09336553  0.1933655  0.2933655  0.3933655
#> 4  04 -1.50373432 -1.4037343 -1.3037343 -1.2037343
#> 5  05  0.34780678  0.4478068  0.5478068  0.6478068
#> 6  06  0.40440574  0.5044057  0.6044057  0.7044057
#> 7  07 -0.66295581 -0.5629558 -0.4629558 -0.3629558
#> 8  08 -0.89928693 -0.7992869 -0.6992869 -0.5992869
#> 9  09  1.55523569  1.6552357  1.7552357  1.8552357
#> 10 10  1.48448713  1.5844871  1.6844871  1.7844871

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.9598706   0.9598706  1.05987062   1.2598706   1.8598706
#> 2  02  1.0623491   1.0623491  1.16234907   1.3623491   1.9623491
#> 3  03 -0.6317978  -0.6317978 -0.53179776  -0.3317978   0.2682022
#> 4  04  1.8233835   1.8233835  1.92338354   2.1233835   2.7233835
#> 5  05 -0.5697667  -0.5697667 -0.46976670  -0.2697667   0.3302333
#> 6  06  1.1341692   1.1341692  1.23416917   1.4341692   2.0341692
#> 7  07 -0.4925014  -0.4925014 -0.39250140  -0.1925014   0.4074986
#> 8  08 -1.9649372  -1.9649372 -1.86493724  -1.6649372  -1.0649372
#> 9  09 -0.3196619  -0.3196619 -0.21966190  -0.0196619   0.5803381
#> 10 10 -0.1241454  -0.1241454 -0.02414541   0.1758546   0.7758546