Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.68498372 -0.68498372 -0.5849837
#> 2  02  0.01583815  0.01583815  0.1158381
#> 3  03  0.55615198  0.55615198  0.6561520
#> 4  04 -0.54315948 -0.54315948 -0.4431595
#> 5  05 -0.15896910 -0.15896910 -0.0589691
#> 6  06 -0.38361653 -0.38361653 -0.2836165
#> 7  07 -2.32364646 -2.32364646 -2.2236465
#> 8  08  0.88275309  0.88275309  0.9827531
#> 9  09  0.80130476  0.80130476  0.9013048
#> 10 10  0.54006259  0.54006259  0.6400626

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.52578066  0.52578066  0.62578066
#> 2  02 -0.13450235 -0.13450235 -0.03450235
#> 3  03  0.33104496  0.33104496  0.43104496
#> 4  04  2.11021234  2.11021234  2.21021234
#> 5  05  1.33963382  1.33963382  1.43963382
#> 6  06 -0.04224050 -0.04224050  0.05775950
#> 7  07 -0.73098092 -0.73098092 -0.63098092
#> 8  08 -1.00949699 -1.00949699 -0.90949699
#> 9  09 -1.09106502 -1.09106502 -0.99106502
#> 10 10  0.01596906  0.01596906  0.11596906

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.66553182  0.76553182  0.8655318  0.96553182
#> 2  02  1.51978507  1.61978507  1.7197851  1.81978507
#> 3  03 -0.27174940 -0.17174940 -0.0717494  0.02825060
#> 4  04  0.96990817  1.06990817  1.1699082  1.26990817
#> 5  05 -0.82212311 -0.72212311 -0.6221231 -0.52212311
#> 6  06 -1.12220576 -1.02220576 -0.9222058 -0.82220576
#> 7  07 -0.94217271 -0.84217271 -0.7421727 -0.64217271
#> 8  08 -0.36857568 -0.26857568 -0.1685757 -0.06857568
#> 9  09 -0.36164713 -0.26164713 -0.1616471 -0.06164713
#> 10 10 -0.06313409  0.03686591  0.1368659  0.23686591

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.21993954  1.21993954   1.3199395  1.519939541   2.1199395
#> 2  02  0.10225690  0.10225690   0.2022569  0.402256905   1.0022569
#> 3  03  1.53109102  1.53109102   1.6310910  1.831091024   2.4310910
#> 4  04  0.92251503  0.92251503   1.0225150  1.222515034   1.8225150
#> 5  05 -0.29172542 -0.29172542  -0.1917254  0.008274578   0.6082746
#> 6  06 -0.27300415 -0.27300415  -0.1730042  0.026995846   0.6269958
#> 7  07 -1.85907795 -1.85907795  -1.7590780 -1.559077954  -0.9590780
#> 8  08  1.82397866  1.82397866   1.9239787  2.123978660   2.7239787
#> 9  09 -1.35439247 -1.35439247  -1.2543925 -1.054392467  -0.4543925
#> 10 10  0.08117057  0.08117057   0.1811706  0.381170573   0.9811706