Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.61014575  0.61014575  0.71014575
#> 2  02  2.10050299  2.10050299  2.20050299
#> 3  03  0.29907019  0.29907019  0.39907019
#> 4  04  0.03348662  0.03348662  0.13348662
#> 5  05 -0.07778739 -0.07778739  0.02221261
#> 6  06  0.02875510  0.02875510  0.12875510
#> 7  07 -1.08266341 -1.08266341 -0.98266341
#> 8  08 -0.68530270 -0.68530270 -0.58530270
#> 9  09  0.87096611  0.87096611  0.97096611
#> 10 10  0.54156452  0.54156452  0.64156452

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.59589481 -0.59589481 -0.4958948
#> 2  02  0.42326775  0.42326775  0.5232678
#> 3  03  0.62438924  0.62438924  0.7243892
#> 4  04  1.58296537  1.58296537  1.6829654
#> 5  05 -0.30105758 -0.30105758 -0.2010576
#> 6  06 -0.29834018 -0.29834018 -0.1983402
#> 7  07  0.35463258  0.35463258  0.4546326
#> 8  08 -1.30267381 -1.30267381 -1.2026738
#> 9  09  0.09553264  0.09553264  0.1955326
#> 10 10  0.99334846  0.99334846  1.0933485

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.6455932 -0.5455932 -0.4455932 -0.3455932
#> 2  02 -0.5277665 -0.4277665 -0.3277665 -0.2277665
#> 3  03  0.1239860  0.2239860  0.3239860  0.4239860
#> 4  04 -1.1187201 -1.0187201 -0.9187201 -0.8187201
#> 5  05 -0.6570350 -0.5570350 -0.4570350 -0.3570350
#> 6  06 -0.5417242 -0.4417242 -0.3417242 -0.2417242
#> 7  07 -0.9450177 -0.8450177 -0.7450177 -0.6450177
#> 8  08  0.4868537  0.5868537  0.6868537  0.7868537
#> 9  09  0.6190482  0.7190482  0.8190482  0.9190482
#> 10 10  0.6006724  0.7006724  0.8006724  0.9006724

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5495961   0.5495961  0.64959607  0.84959607   1.4495961
#> 2  02 -0.7129486  -0.7129486 -0.61294860 -0.41294860   0.1870514
#> 3  03 -0.3180460  -0.3180460 -0.21804596 -0.01804596   0.5819540
#> 4  04  0.1431468   0.1431468  0.24314678  0.44314678   1.0431468
#> 5  05  0.5528720   0.5528720  0.65287199  0.85287199   1.4528720
#> 6  06 -1.2326256  -1.2326256 -1.13262563 -0.93262563  -0.3326256
#> 7  07 -1.2838698  -1.2838698 -1.18386978 -0.98386978  -0.3838698
#> 8  08 -1.3865246  -1.3865246 -1.28652459 -1.08652459  -0.4865246
#> 9  09 -0.1780120  -0.1780120 -0.07801197  0.12198803   0.7219880
#> 10 10  1.5782930   1.5782930  1.67829296  1.87829296   2.4782930