Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1085791 -1.1085791 -1.0085791
#> 2  02  0.3261418  0.3261418  0.4261418
#> 3  03 -1.1467417 -1.1467417 -1.0467417
#> 4  04  0.5376775  0.5376775  0.6376775
#> 5  05  0.5073144  0.5073144  0.6073144
#> 6  06 -0.5057430 -0.5057430 -0.4057430
#> 7  07  0.9016731  0.9016731  1.0016731
#> 8  08 -0.0676568 -0.0676568  0.0323432
#> 9  09  2.0905336  2.0905336  2.1905336
#> 10 10 -0.9986149 -0.9986149 -0.8986149

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.01454995  0.01454995  0.1145499
#> 2  02 -0.82289901 -0.82289901 -0.7228990
#> 3  03  1.30035466  1.30035466  1.4003547
#> 4  04  1.21473664  1.21473664  1.3147366
#> 5  05 -0.55283451 -0.55283451 -0.4528345
#> 6  06 -1.17488162 -1.17488162 -1.0748816
#> 7  07 -0.43004968 -0.43004968 -0.3300497
#> 8  08  0.44607271  0.44607271  0.5460727
#> 9  09 -2.33203397 -2.33203397 -2.2320340
#> 10 10  1.01112033  1.01112033  1.1111203

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1        Y_Z_2       Y_Z_3
#> 1  01  1.29897694  1.39897694  1.498976939  1.59897694
#> 2  02  0.76576296  0.86576296  0.965762964  1.06576296
#> 3  03 -0.14652254 -0.04652254  0.053477455  0.15347746
#> 4  04 -0.06542406  0.03457594  0.134575939  0.23457594
#> 5  05 -0.07415212  0.02584788  0.125847876  0.22584788
#> 6  06 -0.50003026 -0.40003026 -0.300030263 -0.20003026
#> 7  07 -1.67950710 -1.57950710 -1.479507095 -1.37950710
#> 8  08 -0.20129207 -0.10129207 -0.001292075  0.09870793
#> 9  09  1.10082574  1.20082574  1.300825736  1.40082574
#> 10 10  0.44395276  0.54395276  0.643952756  0.74395276

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -2.2908291  -2.2908291  -2.1908291 -1.99082913  -1.3908291
#> 2  02  0.2308739   0.2308739   0.3308739  0.53087391   1.1308739
#> 3  03  1.3199213   1.3199213   1.4199213  1.61992127   2.2199213
#> 4  04 -0.3433582  -0.3433582  -0.2433582 -0.04335824   0.5566418
#> 5  05 -1.4078330  -1.4078330  -1.3078330 -1.10783299  -0.5078330
#> 6  06 -1.0785487  -1.0785487  -0.9785487 -0.77854869  -0.1785487
#> 7  07  0.9276035   0.9276035   1.0276035  1.22760349   1.8276035
#> 8  08  0.4062751   0.4062751   0.5062751  0.70627512   1.3062751
#> 9  09  0.5308327   0.5308327   0.6308327  0.83083267   1.4308327
#> 10 10  0.2164471   0.2164471   0.3164471  0.51644714   1.1164471