Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.2784223  1.2784223  1.37842234
#> 2  02 -0.6620353 -0.6620353 -0.56203535
#> 3  03  1.9525483  1.9525483  2.05254829
#> 4  04 -0.5981500 -0.5981500 -0.49815002
#> 5  05  3.1002315  3.1002315  3.20023146
#> 6  06 -0.9022574 -0.9022574 -0.80225736
#> 7  07  0.7728810  0.7728810  0.87288099
#> 8  08 -0.1666917 -0.1666917 -0.06669166
#> 9  09 -0.2969098 -0.2969098 -0.19690978
#> 10 10  0.5407058  0.5407058  0.64070581

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01  0.840401862  0.840401862  0.9404019
#> 2  02  0.003453507  0.003453507  0.1034535
#> 3  03 -0.013572599 -0.013572599  0.0864274
#> 4  04  1.406623066  1.406623066  1.5066231
#> 5  05 -0.968543865 -0.968543865 -0.8685439
#> 6  06  0.607278787  0.607278787  0.7072788
#> 7  07 -0.920666755 -0.920666755 -0.8206668
#> 8  08  0.107785024  0.107785024  0.2077850
#> 9  09  0.566121843  0.566121843  0.6661218
#> 10 10 -0.231247427 -0.231247427 -0.1312474

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1        Y_Z_2      Y_Z_3
#> 1  01  0.3355070  0.43550702  0.535507017  0.6355070
#> 2  02 -2.0860950 -1.98609502 -1.886095022 -1.7860950
#> 3  03  1.6662341  1.76623413  1.866234129  1.9662341
#> 4  04 -0.1402606 -0.04026055  0.059739448  0.1597394
#> 5  05 -0.1932768 -0.09327679  0.006723215  0.1067232
#> 6  06  0.4149045  0.51490450  0.614904498  0.7149045
#> 7  07 -0.1312571 -0.03125710  0.068742897  0.1687429
#> 8  08  0.2954384  0.39543842  0.495438416  0.5954384
#> 9  09 -0.1606238 -0.06062383  0.039376166  0.1393762
#> 10 10 -1.0493791 -0.94937907 -0.849379067 -0.7493791

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.15905675  1.15905675   1.2590567  1.45905675  2.05905675
#> 2  02  0.71930303  0.71930303   0.8193030  1.01930303  1.61930303
#> 3  03  0.66814980  0.66814980   0.7681498  0.96814980  1.56814980
#> 4  04 -0.85775494 -0.85775494  -0.7577549 -0.55775494  0.04224506
#> 5  05 -2.28006292 -2.28006292  -2.1800629 -1.98006292 -1.38006292
#> 6  06 -1.20512726 -1.20512726  -1.1051273 -0.90512726 -0.30512726
#> 7  07 -0.39560311 -0.39560311  -0.2956031 -0.09560311  0.50439689
#> 8  08  0.07633732  0.07633732   0.1763373  0.37633732  0.97633732
#> 9  09  0.84982721  0.84982721   0.9498272  1.14982721  1.74982721
#> 10 10 -0.56193937 -0.56193937  -0.4619394 -0.26193937  0.33806063