Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.92812226  0.92812226  1.02812226
#> 2  02  0.64578635  0.64578635  0.74578635
#> 3  03 -1.15471719 -1.15471719 -1.05471719
#> 4  04 -0.01547489 -0.01547489  0.08452511
#> 5  05 -0.42896893 -0.42896893 -0.32896893
#> 6  06 -0.70006496 -0.70006496 -0.60006496
#> 7  07 -0.16290911 -0.16290911 -0.06290911
#> 8  08  1.89304213  1.89304213  1.99304213
#> 9  09  0.85468670  0.85468670  0.95468670
#> 10 10 -0.80683918 -0.80683918 -0.70683918

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0        Y_Z_1
#> 1  01 -0.6638796 -0.6638796 -0.563879639
#> 2  02 -0.6203983 -0.6203983 -0.520398294
#> 3  03  1.7610262  1.7610262  1.861026198
#> 4  04  0.4319483  0.4319483  0.531948336
#> 5  05 -1.5413022 -1.5413022 -1.441302246
#> 6  06  1.2890426  1.2890426  1.389042593
#> 7  07 -0.1199585 -0.1199585 -0.019958466
#> 8  08 -1.7055288 -1.7055288 -1.605528819
#> 9  09 -0.1057300 -0.1057300 -0.005730021
#> 10 10  0.7889519  0.7889519  0.888951866

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.2557118 -0.1557118 -0.0557118  0.0442882
#> 2  02  2.2617978  2.3617978  2.4617978  2.5617978
#> 3  03  0.8703325  0.9703325  1.0703325  1.1703325
#> 4  04  0.5044673  0.6044673  0.7044673  0.8044673
#> 5  05 -1.2119149 -1.1119149 -1.0119149 -0.9119149
#> 6  06 -0.6561639 -0.5561639 -0.4561639 -0.3561639
#> 7  07 -2.0653404 -1.9653404 -1.8653404 -1.7653404
#> 8  08  0.9041288  1.0041288  1.1041288  1.2041288
#> 9  09  0.9820684  1.0820684  1.1820684  1.2820684
#> 10 10  0.2152294  0.3152294  0.4152294  0.5152294

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.2253184   1.2253184   1.3253184  1.52531842   2.1253184
#> 2  02  0.4271806   0.4271806   0.5271806  0.72718061   1.3271806
#> 3  03 -0.3188501  -0.3188501  -0.2188501 -0.01885009   0.5811499
#> 4  04 -1.0293982  -1.0293982  -0.9293982 -0.72939818  -0.1293982
#> 5  05  1.1933131   1.1933131   1.2933131  1.49331312   2.0933131
#> 6  06  0.9708344   0.9708344   1.0708344  1.27083436   1.8708344
#> 7  07 -1.3128715  -1.3128715  -1.2128715 -1.01287147  -0.4128715
#> 8  08 -1.7162787  -1.7162787  -1.6162787 -1.41627869  -0.8162787
#> 9  09  0.7084640   0.7084640   0.8084640  1.00846396   1.6084640
#> 10 10  1.0387874   1.0387874   1.1387874  1.33878742   1.9387874