Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.7081591 -0.7081591 -0.6081591
#> 2  02  0.3979719  0.3979719  0.4979719
#> 3  03  0.2339832  0.2339832  0.3339832
#> 4  04  0.1813876  0.1813876  0.2813876
#> 5  05  0.2669916  0.2669916  0.3669916
#> 6  06  1.0235739  1.0235739  1.1235739
#> 7  07  0.0891828  0.0891828  0.1891828
#> 8  08 -1.6086587 -1.6086587 -1.5086587
#> 9  09 -0.3472513 -0.3472513 -0.2472513
#> 10 10 -0.7992814 -0.7992814 -0.6992814

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -2.8192658 -2.8192658 -2.7192658
#> 2  02  0.1835711  0.1835711  0.2835711
#> 3  03 -0.9349094 -0.9349094 -0.8349094
#> 4  04 -0.2543215 -0.2543215 -0.1543215
#> 5  05  0.6115895  0.6115895  0.7115895
#> 6  06 -1.0198129 -1.0198129 -0.9198129
#> 7  07  0.2160377  0.2160377  0.3160377
#> 8  08  0.8089303  0.8089303  0.9089303
#> 9  09 -0.8486540 -0.8486540 -0.7486540
#> 10 10  0.3867247  0.3867247  0.4867247

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.100831939 -1.0008319 -0.9008319 -0.8008319
#> 2  02 -0.954495173 -0.8544952 -0.7544952 -0.6544952
#> 3  03  0.291972367  0.3919724  0.4919724  0.5919724
#> 4  04  0.004690745  0.1046907  0.2046907  0.3046907
#> 5  05  0.280111373  0.3801114  0.4801114  0.5801114
#> 6  06  1.032601780  1.1326018  1.2326018  1.3326018
#> 7  07  0.215806582  0.3158066  0.4158066  0.5158066
#> 8  08  0.140631568  0.2406316  0.3406316  0.4406316
#> 9  09 -2.401595076 -2.3015951 -2.2015951 -2.1015951
#> 10 10 -0.576107762 -0.4761078 -0.3761078 -0.2761078

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.33751035 -0.33751035 -0.23751035 -0.03751035  0.56248965
#> 2  02 -0.17636856 -0.17636856 -0.07636856  0.12363144  0.72363144
#> 3  03 -0.41179420 -0.41179420 -0.31179420 -0.11179420  0.48820580
#> 4  04  0.91184024  0.91184024  1.01184024  1.21184024  1.81184024
#> 5  05  2.11662146  2.11662146  2.21662146  2.41662146  3.01662146
#> 6  06 -1.08004366 -1.08004366 -0.98004366 -0.78004366 -0.18004366
#> 7  07  0.05094447  0.05094447  0.15094447  0.35094447  0.95094447
#> 8  08  0.97850693  0.97850693  1.07850693  1.27850693  1.87850693
#> 9  09 -0.83297421 -0.83297421 -0.73297421 -0.53297421  0.06702579
#> 10 10 -2.07315276 -2.07315276 -1.97315276 -1.77315276 -1.17315276