Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.7680139  0.7680139  0.8680139
#> 2  02  0.7402625  0.7402625  0.8402625
#> 3  03 -0.4963643 -0.4963643 -0.3963643
#> 4  04 -0.7861261 -0.7861261 -0.6861261
#> 5  05  0.7681505  0.7681505  0.8681505
#> 6  06  0.3592363  0.3592363  0.4592363
#> 7  07 -1.0110070 -1.0110070 -0.9110070
#> 8  08  0.7144648  0.7144648  0.8144648
#> 9  09  1.0860778  1.0860778  1.1860778
#> 10 10 -0.9478216 -0.9478216 -0.8478216

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.3808569 -0.3808569 -0.2808569
#> 2  02 -0.7837948 -0.7837948 -0.6837948
#> 3  03  0.5623174  0.5623174  0.6623174
#> 4  04 -0.3003213 -0.3003213 -0.2003213
#> 5  05  0.4113272  0.4113272  0.5113272
#> 6  06 -0.3001798 -0.3001798 -0.2001798
#> 7  07 -1.2833204 -1.2833204 -1.1833204
#> 8  08 -0.6060568 -0.6060568 -0.5060568
#> 9  09  0.1481620  0.1481620  0.2481620
#> 10 10  1.0619233  1.0619233  1.1619233

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.5266624 -0.42666238 -0.32666238 -0.22666238
#> 2  02  0.0964011  0.19640110  0.29640110  0.39640110
#> 3  03  0.3293215  0.42932150  0.52932150  0.62932150
#> 4  04  0.6207155  0.72071551  0.82071551  0.92071551
#> 5  05 -1.3713360 -1.27133604 -1.17133604 -1.07133604
#> 6  06 -0.3368583 -0.23685825 -0.13685825 -0.03685825
#> 7  07  0.5715930  0.67159295  0.77159295  0.87159295
#> 8  08  0.1616502  0.26165017  0.36165017  0.46165017
#> 9  09 -0.1137894 -0.01378941  0.08621059  0.18621059
#> 10 10 -0.4232124 -0.32321235 -0.22321235 -0.12321235

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0   Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.09187472  0.09187472  0.19187472  0.3918747230 0.991874723
#> 2  02  1.30001374  1.30001374  1.40001374  1.6000137426 2.200013743
#> 3  03 -0.89042840 -0.89042840 -0.79042840 -0.5904283969 0.009571603
#> 4  04 -0.41077670 -0.41077670 -0.31077670 -0.1107766974 0.489223303
#> 5  05  0.04413359  0.04413359  0.14413359  0.3441335851 0.944133585
#> 6  06  1.26399001  1.26399001  1.36399001  1.5639900149 2.163990015
#> 7  07  0.26963604  0.26963604  0.36963604  0.5696360408 1.169636041
#> 8  08 -0.02482487 -0.02482487  0.07517513  0.2751751292 0.875175129
#> 9  09 -0.30079568 -0.30079568 -0.20079568 -0.0007956813 0.599204319
#> 10 10  0.24101050  0.24101050  0.34101050  0.5410105018 1.141010502