Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID             U         Y_Z_0       Y_Z_1
#> 1  01 -0.3923418867 -0.3923418867 -0.29234189
#> 2  02 -0.5160102160 -0.5160102160 -0.41601022
#> 3  03 -0.9988233930 -0.9988233930 -0.89882339
#> 4  04  1.0163498065  1.0163498065  1.11634981
#> 5  05 -1.8136046685 -1.8136046685 -1.71360467
#> 6  06 -1.1574534636 -1.1574534636 -1.05745346
#> 7  07 -0.0744019855 -0.0744019855  0.02559801
#> 8  08 -0.8579220753 -0.8579220753 -0.75792208
#> 9  09  0.0005441256  0.0005441256  0.10054413
#> 10 10 -1.3179449591 -1.3179449591 -1.21794496

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.0637818 -1.0637818 -0.9637818
#> 2  02 -0.3040027 -0.3040027 -0.2040027
#> 3  03  1.2010790  1.2010790  1.3010790
#> 4  04  0.9611892  0.9611892  1.0611892
#> 5  05  0.8915127  0.8915127  0.9915127
#> 6  06  2.7147920  2.7147920  2.8147920
#> 7  07  0.1554174  0.1554174  0.2554174
#> 8  08  1.4558342  1.4558342  1.5558342
#> 9  09 -0.5488781 -0.5488781 -0.4488781
#> 10 10  0.9026010  0.9026010  1.0026010

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.676134729  0.77613473  0.8761347  0.9761347
#> 2  02 -0.792345080 -0.69234508 -0.5923451 -0.4923451
#> 3  03  1.471352110  1.57135211  1.6713521  1.7713521
#> 4  04 -0.001286683  0.09871332  0.1987133  0.2987133
#> 5  05 -0.907448208 -0.80744821 -0.7074482 -0.6074482
#> 6  06  0.103604427  0.20360443  0.3036044  0.4036044
#> 7  07 -2.333275083 -2.23327508 -2.1332751 -2.0332751
#> 8  08  1.044009252  1.14400925  1.2440093  1.3440093
#> 9  09 -1.480267398 -1.38026740 -1.2802674 -1.1802674
#> 10 10  0.580376479  0.68037648  0.7803765  0.8803765

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5239791   0.5239791   0.6239791  0.823979120   1.4239791
#> 2  02  0.8307454   0.8307454   0.9307454  1.130745449   1.7307454
#> 3  03 -0.6232065  -0.6232065  -0.5232065 -0.323206528   0.2767935
#> 4  04 -0.5704384  -0.5704384  -0.4704384 -0.270438361   0.3295616
#> 5  05 -0.3012012  -0.3012012  -0.2012012 -0.001201214   0.5987988
#> 6  06 -1.3638394  -1.3638394  -1.2638394 -1.063839428  -0.4638394
#> 7  07  0.2403224   0.2403224   0.3403224  0.540322443   1.1403224
#> 8  08  1.8381008   1.8381008   1.9381008  2.138100825   2.7381008
#> 9  09 -0.4556329  -0.4556329  -0.3556329 -0.155632864   0.4443671
#> 10 10  0.8672199   0.8672199   0.9672199  1.167219897   1.7672199