Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.2760967  0.2760967  0.37609674
#> 2  02 -0.9540323 -0.9540323 -0.85403227
#> 3  03  1.5326727  1.5326727  1.63267266
#> 4  04  0.9661987  0.9661987  1.06619872
#> 5  05 -0.9734825 -0.9734825 -0.87348252
#> 6  06 -0.5585692 -0.5585692 -0.45856925
#> 7  07 -0.1606536 -0.1606536 -0.06065357
#> 8  08 -0.5909818 -0.5909818 -0.49098180
#> 9  09 -1.6782951 -1.6782951 -1.57829513
#> 10 10 -0.1694169 -0.1694169 -0.06941690

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.8267809 -1.8267809 -1.7267809
#> 2  02  0.4837199  0.4837199  0.5837199
#> 3  03 -0.8705655 -0.8705655 -0.7705655
#> 4  04 -0.1340596 -0.1340596 -0.0340596
#> 5  05  2.3069656  2.3069656  2.4069656
#> 6  06  0.2105458  0.2105458  0.3105458
#> 7  07  0.6340025  0.6340025  0.7340025
#> 8  08  0.1072994  0.1072994  0.2072994
#> 9  09 -1.1005340 -1.1005340 -1.0005340
#> 10 10  0.3588731  0.3588731  0.4588731

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  1.4960942  1.59609418  1.69609418  1.7960942
#> 2  02  0.5697956  0.66979564  0.76979564  0.8697956
#> 3  03  0.4772174  0.57721739  0.67721739  0.7772174
#> 4  04 -0.1774418 -0.07744182  0.02255818  0.1225582
#> 5  05  0.2391678  0.33916776  0.43916776  0.5391678
#> 6  06 -1.7689006 -1.66890056 -1.56890056 -1.4689006
#> 7  07 -1.6002884 -1.50028838 -1.40028838 -1.3002884
#> 8  08 -1.1905643 -1.09056426 -0.99056426 -0.8905643
#> 9  09  0.4209025  0.52090250  0.62090250  0.7209025
#> 10 10  0.3253088  0.42530877  0.52530877  0.6253088

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.8151910  -0.8151910  -0.7151910 -0.51519105  0.08480895
#> 2  02 -1.0661037  -1.0661037  -0.9661037 -0.76610372 -0.16610372
#> 3  03  0.7541621   0.7541621   0.8541621  1.05416212  1.65416212
#> 4  04  0.3091471   0.3091471   0.4091471  0.60914710  1.20914710
#> 5  05 -0.7272036  -0.7272036  -0.6272036 -0.42720365  0.17279635
#> 6  06  1.4853810   1.4853810   1.5853810  1.78538096  2.38538096
#> 7  07  0.7165759   0.7165759   0.8165759  1.01657590  1.61657590
#> 8  08 -0.3135608  -0.3135608  -0.2135608 -0.01356078  0.58643922
#> 9  09 -0.5487764  -0.5487764  -0.4487764 -0.24877643  0.35122357
#> 10 10 -0.2397890  -0.2397890  -0.1397890  0.06021103  0.66021103