Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.02586477  0.02586477  0.1258648
#> 2  02 -2.21541797 -2.21541797 -2.1154180
#> 3  03  1.49568711  1.49568711  1.5956871
#> 4  04  0.64411507  0.64411507  0.7441151
#> 5  05  1.29154928  1.29154928  1.3915493
#> 6  06  1.24374654  1.24374654  1.3437465
#> 7  07  0.21001010  0.21001010  0.3100101
#> 8  08 -0.47428886 -0.47428886 -0.3742889
#> 9  09  1.30094370  1.30094370  1.4009437
#> 10 10 -0.97011887 -0.97011887 -0.8701189

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.29552628 -0.29552628 -0.1955263
#> 2  02  0.09174067  0.09174067  0.1917407
#> 3  03 -1.06754779 -1.06754779 -0.9675478
#> 4  04  0.05833092  0.05833092  0.1583309
#> 5  05  0.23778985  0.23778985  0.3377898
#> 6  06  0.72709319  0.72709319  0.8270932
#> 7  07  0.57543681  0.57543681  0.6754368
#> 8  08  0.19133497  0.19133497  0.2913350
#> 9  09 -0.53879242 -0.53879242 -0.4387924
#> 10 10 -1.40302249 -1.40302249 -1.3030225

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1        Y_Z_2       Y_Z_3
#> 1  01 -1.0683756 -0.96837559 -0.868375591 -0.76837559
#> 2  02 -0.1909803 -0.09098035  0.009019654  0.10901965
#> 3  03 -1.5082751 -1.40827506 -1.308275056 -1.20827506
#> 4  04  0.5177219  0.61772193  0.717721926  0.81772193
#> 5  05 -0.2555706 -0.15557063 -0.055570628  0.04442937
#> 6  06  2.2060030  2.30600296  2.406002962  2.50600296
#> 7  07  0.1185485  0.21854850  0.318548505  0.41854850
#> 8  08 -1.0792767 -0.97927668 -0.879276683 -0.77927668
#> 9  09 -0.4442739 -0.34427389 -0.244273887 -0.14427389
#> 10 10  1.2832469  1.38324694  1.483246938  1.58324694

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.7049497  -0.7049497  -0.6049497  -0.4049497  0.19505025
#> 2  02 -0.4890633  -0.4890633  -0.3890633  -0.1890633  0.41093671
#> 3  03  2.4604613   2.4604613   2.5604613   2.7604613  3.36046128
#> 4  04  0.5911599   0.5911599   0.6911599   0.8911599  1.49115992
#> 5  05  0.1481359   0.1481359   0.2481359   0.4481359  1.04813590
#> 6  06  1.1067808   1.1067808   1.2067808   1.4067808  2.00678080
#> 7  07 -0.8059231  -0.8059231  -0.7059231  -0.5059231  0.09407685
#> 8  08  1.1033445   1.1033445   1.2033445   1.4033445  2.00334453
#> 9  09  0.3640282   0.3640282   0.4640282   0.6640282  1.26402818
#> 10 10  0.6619098   0.6619098   0.7619098   0.9619098  1.56190983