Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.9856954 -0.9856954 -0.8856954
#> 2  02 -0.8477302 -0.8477302 -0.7477302
#> 3  03 -0.4237819 -0.4237819 -0.3237819
#> 4  04 -0.9640272 -0.9640272 -0.8640272
#> 5  05  0.2860357  0.2860357  0.3860357
#> 6  06  0.7767212  0.7767212  0.8767212
#> 7  07 -0.6833906 -0.6833906 -0.5833906
#> 8  08 -1.4010734 -1.4010734 -1.3010734
#> 9  09 -0.2714169 -0.2714169 -0.1714169
#> 10 10  0.4451608  0.4451608  0.5451608

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.07716539  0.07716539  0.1771654
#> 2  02 -1.88969533 -1.88969533 -1.7896953
#> 3  03  0.07369645  0.07369645  0.1736965
#> 4  04  1.19083937  1.19083937  1.2908394
#> 5  05  1.10945000  1.10945000  1.2094500
#> 6  06  0.86063995  0.86063995  0.9606400
#> 7  07  1.63029366  1.63029366  1.7302937
#> 8  08  1.68043459  1.68043459  1.7804346
#> 9  09  1.57600326  1.57600326  1.6760033
#> 10 10  0.18374165  0.18374165  0.2837416

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01  1.3005700  1.4005700  1.50057001  1.600570011
#> 2  02 -0.8743810 -0.7743810 -0.67438099 -0.574380993
#> 3  03 -0.4451689 -0.3451689 -0.24516894 -0.145168943
#> 4  04  0.6409037  0.7409037  0.84090366  0.940903660
#> 5  05  0.3745270  0.4745270  0.57452701  0.674527007
#> 6  06  0.8093769  0.9093769  1.00937694  1.109376944
#> 7  07 -1.3851238 -1.2851238 -1.18512385 -1.085123850
#> 8  08  0.1596942  0.2596942  0.35969417  0.459694168
#> 9  09 -0.2961083 -0.1961083 -0.09610835  0.003891651
#> 10 10 -0.6288296 -0.5288296 -0.42882961 -0.328829608

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.3137818   0.3137818  0.41378177  0.61378177   1.2137818
#> 2  02 -0.2474897  -0.2474897 -0.14748974  0.05251026   0.6525103
#> 3  03 -1.2489261  -1.2489261 -1.14892606 -0.94892606  -0.3489261
#> 4  04  1.3841715   1.3841715  1.48417153  1.68417153   2.2841715
#> 5  05  0.9142033   0.9142033  1.01420331  1.21420331   1.8142033
#> 6  06  0.4774785   0.4774785  0.57747847  0.77747847   1.3774785
#> 7  07 -0.1284719  -0.1284719 -0.02847192  0.17152808   0.7715281
#> 8  08  0.7649121   0.7649121  0.86491213  1.06491213   1.6649121
#> 9  09  0.6307846   0.6307846  0.73078463  0.93078463   1.5307846
#> 10 10 -1.1000834  -1.1000834 -1.00008338 -0.80008338  -0.2000834