Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.400962950  0.400962950  0.50096295
#> 2  02  0.789107941  0.789107941  0.88910794
#> 3  03  0.423376126  0.423376126  0.52337613
#> 4  04 -0.127450185 -0.127450185 -0.02745019
#> 5  05 -0.900294145 -0.900294145 -0.80029415
#> 6  06 -0.372189737 -0.372189737 -0.27218974
#> 7  07  1.001784716  1.001784716  1.10178472
#> 8  08  0.007787989  0.007787989  0.10778799
#> 9  09 -1.327100039 -1.327100039 -1.22710004
#> 10 10  1.200573268  1.200573268  1.30057327

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.3504157  0.3504157  0.4504157
#> 2  02  1.2306852  1.2306852  1.3306852
#> 3  03 -0.7486209 -0.7486209 -0.6486209
#> 4  04  1.8142122  1.8142122  1.9142122
#> 5  05  0.8410075  0.8410075  0.9410075
#> 6  06  1.3956823  1.3956823  1.4956823
#> 7  07  0.2991088  0.2991088  0.3991088
#> 8  08  0.5195012  0.5195012  0.6195012
#> 9  09 -1.3874236 -1.3874236 -1.2874236
#> 10 10  0.3581435  0.3581435  0.4581435

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.9882050 -0.88820497 -0.78820497 -0.68820497
#> 2  02  1.2181088  1.31810884  1.41810884  1.51810884
#> 3  03 -1.7781317 -1.67813171 -1.57813171 -1.47813171
#> 4  04  0.9523510  1.05235099  1.15235099  1.25235099
#> 5  05 -0.3631094 -0.26310945 -0.16310945 -0.06310945
#> 6  06 -1.1600342 -1.06003425 -0.96003425 -0.86003425
#> 7  07  1.2785911  1.37859108  1.47859108  1.57859108
#> 8  08 -0.1748035 -0.07480348  0.02519652  0.12519652
#> 9  09  0.4860707  0.58607070  0.68607070  0.78607070
#> 10 10  0.6071007  0.70710069  0.80710069  0.90710069

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.640978471  0.640978471   0.7409785   0.9409785  1.54097847
#> 2  02 -0.994675697 -0.994675697  -0.8946757  -0.6946757 -0.09467570
#> 3  03  0.357203387  0.357203387   0.4572034   0.6572034  1.25720339
#> 4  04 -1.153158137 -1.153158137  -1.0531581  -0.8531581 -0.25315814
#> 5  05 -1.555400309 -1.555400309  -1.4554003  -1.2554003 -0.65540031
#> 6  06  0.826498268  0.826498268   0.9264983   1.1264983  1.72649827
#> 7  07 -0.810087659 -0.810087659  -0.7100877  -0.5100877  0.08991234
#> 8  08 -0.560893698 -0.560893698  -0.4608937  -0.2608937  0.33910630
#> 9  09  0.006393238  0.006393238   0.1063932   0.3063932  0.90639324
#> 10 10 -1.408556157 -1.408556157  -1.3085562  -1.1085562 -0.50855616