Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.33475180  0.33475180  0.43475180
#> 2  02 -0.91625726 -0.91625726 -0.81625726
#> 3  03 -0.05297474 -0.05297474  0.04702526
#> 4  04 -1.40232743 -1.40232743 -1.30232743
#> 5  05 -1.55713798 -1.55713798 -1.45713798
#> 6  06 -1.03378086 -1.03378086 -0.93378086
#> 7  07  0.05588824  0.05588824  0.15588824
#> 8  08 -0.16639356 -0.16639356 -0.06639356
#> 9  09  0.31841111  0.31841111  0.41841111
#> 10 10 -0.16408315 -0.16408315 -0.06408315

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -2.31122894 -2.31122894 -2.21122894
#> 2  02  1.32428767  1.32428767  1.42428767
#> 3  03 -1.49101125 -1.49101125 -1.39101125
#> 4  04 -0.04601687 -0.04601687  0.05398313
#> 5  05 -2.01380641 -2.01380641 -1.91380641
#> 6  06  0.66343035  0.66343035  0.76343035
#> 7  07  0.80825516  0.80825516  0.90825516
#> 8  08 -0.67086801 -0.67086801 -0.57086801
#> 9  09 -0.25215806 -0.25215806 -0.15215806
#> 10 10  0.07948421  0.07948421  0.17948421

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.54414618 -1.44414618 -1.3441462 -1.2441462
#> 2  02  0.14929567  0.24929567  0.3492957  0.4492957
#> 3  03 -1.28211277 -1.18211277 -1.0821128 -0.9821128
#> 4  04  0.05673488  0.15673488  0.2567349  0.3567349
#> 5  05 -0.76119780 -0.66119780 -0.5611978 -0.4611978
#> 6  06 -0.08011025  0.01988975  0.1198897  0.2198897
#> 7  07 -2.83806192 -2.73806192 -2.6380619 -2.5380619
#> 8  08 -0.07356374  0.02643626  0.1264363  0.2264363
#> 9  09  0.76608646  0.86608646  0.9660865  1.0660865
#> 10 10 -0.54320888 -0.44320888 -0.3432089 -0.2432089

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.12651373  0.12651373  0.226513735  0.42651373   1.0265137
#> 2  02 -0.22069626 -0.22069626 -0.120696263  0.07930374   0.6793037
#> 3  03 -1.31554917 -1.31554917 -1.215549171 -1.01554917  -0.4155492
#> 4  04  0.53903374  0.53903374  0.639033737  0.83903374   1.4390337
#> 5  05  0.64774523  0.64774523  0.747745231  0.94774523   1.5477452
#> 6  06 -0.43030740 -0.43030740 -0.330307405 -0.13030740   0.4696926
#> 7  07  0.20000978  0.20000978  0.300009776  0.50000978   1.1000098
#> 8  08 -0.06996198 -0.06996198  0.030038018  0.23003802   0.8300380
#> 9  09 -0.10991611 -0.10991611 -0.009916107  0.19008389   0.7900839
#> 10 10  0.37077745  0.37077745  0.470777448  0.67077745   1.2707774