Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.6589772 -0.6589772 -0.5589772
#> 2  02 -1.1178256 -1.1178256 -1.0178256
#> 3  03  0.9378802  0.9378802  1.0378802
#> 4  04 -0.8834207 -0.8834207 -0.7834207
#> 5  05 -1.0028313 -1.0028313 -0.9028313
#> 6  06 -1.2036538 -1.2036538 -1.1036538
#> 7  07  1.0648569  1.0648569  1.1648569
#> 8  08  0.6785158  0.6785158  0.7785158
#> 9  09 -0.7758330 -0.7758330 -0.6758330
#> 10 10  0.7490912  0.7490912  0.8490912

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.6932576  0.6932576  0.79325755
#> 2  02 -0.2838584 -0.2838584 -0.18385840
#> 3  03 -1.7343603 -1.7343603 -1.63436031
#> 4  04  1.2873989  1.2873989  1.38739889
#> 5  05  0.9908120  0.9908120  1.09081195
#> 6  06 -0.1333553 -0.1333553 -0.03335534
#> 7  07 -0.8156163 -0.8156163 -0.71561634
#> 8  08 -0.3182072 -0.3182072 -0.21820716
#> 9  09 -0.9138126 -0.9138126 -0.81381260
#> 10 10  0.7266640  0.7266640  0.82666399

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.0821306  0.1821306  0.28213060  0.38213060
#> 2  02 -0.6895439 -0.5895439 -0.48954390 -0.38954390
#> 3  03 -0.6924103 -0.5924103 -0.49241033 -0.39241033
#> 4  04  0.8194218  0.9194218  1.01942179  1.11942179
#> 5  05 -0.7672144 -0.6672144 -0.56721437 -0.46721437
#> 6  06  0.8998503  0.9998503  1.09985030  1.19985030
#> 7  07 -0.2870446 -0.1870446 -0.08704458  0.01295542
#> 8  08  2.0969252  2.1969252  2.29692522  2.39692522
#> 9  09 -0.2789399 -0.1789399 -0.07893992  0.02106008
#> 10 10 -0.6938207 -0.5938207 -0.49382074 -0.39382074

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.56753387  1.56753387   1.6675339  1.86753387  2.46753387
#> 2  02 -0.71633707 -0.71633707  -0.6163371 -0.41633707  0.18366293
#> 3  03 -0.95209730 -0.95209730  -0.8520973 -0.65209730 -0.05209730
#> 4  04  1.40434834  1.40434834   1.5043483  1.70434834  2.30434834
#> 5  05  0.71269157  0.71269157   0.8126916  1.01269157  1.61269157
#> 6  06 -0.86864316 -0.86864316  -0.7686432 -0.56864316  0.03135684
#> 7  07 -0.28939231 -0.28939231  -0.1893923  0.01060769  0.61060769
#> 8  08 -1.58166895 -1.58166895  -1.4816690 -1.28166895 -0.68166895
#> 9  09  0.02650092  0.02650092   0.1265009  0.32650092  0.92650092
#> 10 10 -1.34511750 -1.34511750  -1.2451175 -1.04511750 -0.44511750