Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.3477465 -0.3477465 -0.2477465
#> 2  02  1.2339808  1.2339808  1.3339808
#> 3  03  0.2851184  0.2851184  0.3851184
#> 4  04 -0.5625463 -0.5625463 -0.4625463
#> 5  05 -0.5461952 -0.5461952 -0.4461952
#> 6  06 -1.0854901 -1.0854901 -0.9854901
#> 7  07 -0.0485531 -0.0485531  0.0514469
#> 8  08  0.6925576  0.6925576  0.7925576
#> 9  09 -0.4289805 -0.4289805 -0.3289805
#> 10 10  0.7279687  0.7279687  0.8279687

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.90052559 -0.90052559 -0.80052559
#> 2  02 -0.75243590 -0.75243590 -0.65243590
#> 3  03  0.25134671  0.25134671  0.35134671
#> 4  04 -0.71603240 -0.71603240 -0.61603240
#> 5  05 -0.33002298 -0.33002298 -0.23002298
#> 6  06 -1.44880860 -1.44880860 -1.34880860
#> 7  07  1.22616919  1.22616919  1.32616919
#> 8  08  1.49088336  1.49088336  1.59088336
#> 9  09 -0.01703014 -0.01703014  0.08296986
#> 10 10 -0.54951589 -0.54951589 -0.44951589

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.4385589 -0.33855887 -0.23855887 -0.13855887
#> 2  02  2.2415082  2.34150825  2.44150825  2.54150825
#> 3  03 -0.1646420 -0.06464202  0.03535798  0.13535798
#> 4  04 -0.1136037 -0.01360371  0.08639629  0.18639629
#> 5  05 -2.0821909 -1.98219091 -1.88219091 -1.78219091
#> 6  06 -0.0413910  0.05860900  0.15860900  0.25860900
#> 7  07  0.6632495  0.76324953  0.86324953  0.96324953
#> 8  08 -0.2208025 -0.12080253 -0.02080253  0.07919747
#> 9  09 -0.2183560 -0.11835596 -0.01835596  0.08164404
#> 10 10  1.1692721  1.26927206  1.36927206  1.46927206

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.14873558  0.14873558   0.2487356   0.4487356  1.04873558
#> 2  02  0.60094492  0.60094492   0.7009449   0.9009449  1.50094492
#> 3  03  0.50685431  0.50685431   0.6068543   0.8068543  1.40685431
#> 4  04  1.14006023  1.14006023   1.2400602   1.4400602  2.04006023
#> 5  05  0.91405946  0.91405946   1.0140595   1.2140595  1.81405946
#> 6  06 -1.04452910 -1.04452910  -0.9445291  -0.7445291 -0.14452910
#> 7  07 -0.83223223 -0.83223223  -0.7322322  -0.5322322  0.06776777
#> 8  08 -0.54931787 -0.54931787  -0.4493179  -0.2493179  0.35068213
#> 9  09  1.16605550  1.16605550   1.2660555   1.4660555  2.06605550
#> 10 10  0.03509694  0.03509694   0.1350969   0.3350969  0.93509694