Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.50016979  1.50016979  1.6001698
#> 2  02  0.03575591  0.03575591  0.1357559
#> 3  03  0.43051019  0.43051019  0.5305102
#> 4  04  1.47421283  1.47421283  1.5742128
#> 5  05  0.43035703  0.43035703  0.5303570
#> 6  06 -1.75972652 -1.75972652 -1.6597265
#> 7  07 -1.14448593 -1.14448593 -1.0444859
#> 8  08 -2.14263077 -2.14263077 -2.0426308
#> 9  09 -0.31518963 -0.31518963 -0.2151896
#> 10 10  0.70275728  0.70275728  0.8027573

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.485225682  0.485225682  0.58522568
#> 2  02  0.126543365  0.126543365  0.22654336
#> 3  03 -1.362759578 -1.362759578 -1.26275958
#> 4  04 -1.481395130 -1.481395130 -1.38139513
#> 5  05 -0.565012507 -0.565012507 -0.46501251
#> 6  06  0.728451962  0.728451962  0.82845196
#> 7  07 -0.027369289 -0.027369289  0.07263071
#> 8  08 -0.310603077 -0.310603077 -0.21060308
#> 9  09 -0.008017534 -0.008017534  0.09198247
#> 10 10 -0.683233127 -0.683233127 -0.58323313

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2         Y_Z_3
#> 1  01  0.9165614  1.01656139  1.11656139  1.2165613871
#> 2  02  1.2709555  1.37095552  1.47095552  1.5709555210
#> 3  03 -0.1779643 -0.07796432  0.02203568  0.1220356798
#> 4  04 -1.0703717 -0.97037174 -0.87037174 -0.7703717402
#> 5  05  1.5246457  1.62464575  1.72464575  1.8246457478
#> 6  06 -0.3004054 -0.20040536 -0.10040536 -0.0004053556
#> 7  07 -0.8512601 -0.75126011 -0.65126011 -0.5512601053
#> 8  08  1.4627468  1.56274681  1.66274681  1.7627468108
#> 9  09 -0.8009350 -0.70093497 -0.60093497 -0.5009349742
#> 10 10  0.6940721  0.79407211  0.89407211  0.9940721067

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.5359212  -0.5359212  -0.4359212  -0.2359212   0.3640788
#> 2  02 -0.5248648  -0.5248648  -0.4248648  -0.2248648   0.3751352
#> 3  03  1.4183320   1.4183320   1.5183320   1.7183320   2.3183320
#> 4  04  0.2925036   0.2925036   0.3925036   0.5925036   1.1925036
#> 5  05 -1.3347344  -1.3347344  -1.2347344  -1.0347344  -0.4347344
#> 6  06 -1.6580765  -1.6580765  -1.5580765  -1.3580765  -0.7580765
#> 7  07  1.4332480   1.4332480   1.5332480   1.7332480   2.3332480
#> 8  08 -1.1383098  -1.1383098  -1.0383098  -0.8383098  -0.2383098
#> 9  09 -0.7532294  -0.7532294  -0.6532294  -0.4532294   0.1467706
#> 10 10  0.6687627   0.6687627   0.7687627   0.9687627   1.5687627