Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.5155047  1.5155047  1.6155047
#> 2  02 -2.3744711 -2.3744711 -2.2744711
#> 3  03 -0.7355407 -0.7355407 -0.6355407
#> 4  04 -2.0982528 -2.0982528 -1.9982528
#> 5  05  1.3463372  1.3463372  1.4463372
#> 6  06 -0.7447182 -0.7447182 -0.6447182
#> 7  07 -2.0552398 -2.0552398 -1.9552398
#> 8  08 -0.2885988 -0.2885988 -0.1885988
#> 9  09  0.3307324  0.3307324  0.4307324
#> 10 10 -0.7284924 -0.7284924 -0.6284924

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.37225463  0.37225463  0.4722546
#> 2  02  0.01772875  0.01772875  0.1177287
#> 3  03  0.10901202  0.10901202  0.2090120
#> 4  04 -0.20647386 -0.20647386 -0.1064739
#> 5  05 -1.48024788 -1.48024788 -1.3802479
#> 6  06 -1.57518849 -1.57518849 -1.4751885
#> 7  07  0.36187205  0.36187205  0.4618721
#> 8  08 -1.63782107 -1.63782107 -1.5378211
#> 9  09 -1.27196850 -1.27196850 -1.1719685
#> 10 10 -0.47747741 -0.47747741 -0.3774774

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -0.13039657 -0.03039657  0.06960343  0.1696034
#> 2  02  0.94609771  1.04609771  1.14609771  1.2460977
#> 3  03 -0.94886780 -0.84886780 -0.74886780 -0.6488678
#> 4  04 -0.58105530 -0.48105530 -0.38105530 -0.2810553
#> 5  05  0.88076569  0.98076569  1.08076569  1.1807657
#> 6  06  1.06018687  1.16018687  1.26018687  1.3601869
#> 7  07 -1.04347821 -0.94347821 -0.84347821 -0.7434782
#> 8  08 -0.08894237  0.01105763  0.11105763  0.2110576
#> 9  09  0.53020505  0.63020505  0.73020505  0.8302050
#> 10 10 -0.49919203 -0.39919203 -0.29919203 -0.1991920

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1
#> 1  01 -0.005611632 -0.005611632  0.09438837  0.29438837  0.894388368
#> 2  02  1.434669439  1.434669439  1.53466944  1.73466944  2.334669439
#> 3  03  1.163943765  1.163943765  1.26394377  1.46394377  2.063943765
#> 4  04 -0.396965047 -0.396965047 -0.29696505 -0.09696505  0.503034953
#> 5  05  0.098163724  0.098163724  0.19816372  0.39816372  0.998163724
#> 6  06 -0.539992988 -0.539992988 -0.43999299 -0.23999299  0.360007012
#> 7  07  0.654730876  0.654730876  0.75473088  0.95473088  1.554730876
#> 8  08  0.590676305  0.590676305  0.69067630  0.89067630  1.490676305
#> 9  09 -1.005433963 -1.005433963 -0.90543396 -0.70543396 -0.105433963
#> 10 10 -0.891662797 -0.891662797 -0.79166280 -0.59166280  0.008337203