Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01 -0.223164452 -0.223164452 -0.1231645
#> 2  02  0.003888752  0.003888752  0.1038888
#> 3  03 -1.675031057 -1.675031057 -1.5750311
#> 4  04  0.726928512  0.726928512  0.8269285
#> 5  05  0.587774378  0.587774378  0.6877744
#> 6  06  0.977713176  0.977713176  1.0777132
#> 7  07  0.550937886  0.550937886  0.6509379
#> 8  08  0.112936825  0.112936825  0.2129368
#> 9  09  0.056126350  0.056126350  0.1561264
#> 10 10 -1.916924592 -1.916924592 -1.8169246

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.68673510  1.68673510  1.7867351
#> 2  02 -1.38257509 -1.38257509 -1.2825751
#> 3  03  0.41191636  0.41191636  0.5119164
#> 4  04 -1.04546012 -1.04546012 -0.9454601
#> 5  05  1.10021394  1.10021394  1.2002139
#> 6  06 -0.97490053 -0.97490053 -0.8749005
#> 7  07  0.09806103  0.09806103  0.1980610
#> 8  08  1.13888610  1.13888610  1.2388861
#> 9  09  0.57255101  0.57255101  0.6725510
#> 10 10 -2.20067068 -2.20067068 -2.1006707

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  1.11574377  1.2157438  1.3157438  1.4157438
#> 2  02 -0.18256570 -0.0825657  0.0174343  0.1174343
#> 3  03 -0.65227841 -0.5522784 -0.4522784 -0.3522784
#> 4  04  0.22101854  0.3210185  0.4210185  0.5210185
#> 5  05  1.11887217  1.2188722  1.3188722  1.4188722
#> 6  06  0.69546892  0.7954689  0.8954689  0.9954689
#> 7  07  0.07782617  0.1778262  0.2778262  0.3778262
#> 8  08  0.11333951  0.2133395  0.3133395  0.4133395
#> 9  09  0.60371578  0.7037158  0.8037158  0.9037158
#> 10 10 -0.46439057 -0.3643906 -0.2643906 -0.1643906

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.575054581 -0.575054581 -0.47505458  -0.2750546   0.3249454
#> 2  02  0.320823113  0.320823113  0.42082311   0.6208231   1.2208231
#> 3  03  0.156706707  0.156706707  0.25670671   0.4567067   1.0567067
#> 4  04  0.462052410  0.462052410  0.56205241   0.7620524   1.3620524
#> 5  05  1.077890750  1.077890750  1.17789075   1.3778907   1.9778907
#> 6  06  0.219124221  0.219124221  0.31912422   0.5191242   1.1191242
#> 7  07  0.974941263  0.974941263  1.07494126   1.2749413   1.8749413
#> 8  08  1.421157063  1.421157063  1.52115706   1.7211571   2.3211571
#> 9  09 -0.057569072 -0.057569072  0.04243093   0.2424309   0.8424309
#> 10 10  0.006198333  0.006198333  0.10619833   0.3061983   0.9061983