Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.3683041 -0.3683041 -0.2683041
#> 2  02 -0.5769767 -0.5769767 -0.4769767
#> 3  03  0.6861038  0.6861038  0.7861038
#> 4  04  0.9237798  0.9237798  1.0237798
#> 5  05  0.3671458  0.3671458  0.4671458
#> 6  06 -0.8981993 -0.8981993 -0.7981993
#> 7  07 -1.4879750 -1.4879750 -1.3879750
#> 8  08  0.6959410  0.6959410  0.7959410
#> 9  09  0.8956242  0.8956242  0.9956242
#> 10 10  0.5371665  0.5371665  0.6371665

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.7981501  0.7981501  0.8981501
#> 2  02  0.2834591  0.2834591  0.3834591
#> 3  03 -1.6492140 -1.6492140 -1.5492140
#> 4  04  0.6453870  0.6453870  0.7453870
#> 5  05 -1.0600483 -1.0600483 -0.9600483
#> 6  06  1.2322284  1.2322284  1.3322284
#> 7  07  1.0908071  1.0908071  1.1908071
#> 8  08 -0.9727254 -0.9727254 -0.8727254
#> 9  09 -0.7759122 -0.7759122 -0.6759122
#> 10 10  2.2081838  2.2081838  2.3081838

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01 -1.2338057 -1.1338057 -1.03380570 -0.933805705
#> 2  02  0.8369121  0.9369121  1.03691212  1.136912119
#> 3  03  0.6810272  0.7810272  0.88102719  0.981027194
#> 4  04  0.6845331  0.7845331  0.88453311  0.984533107
#> 5  05 -0.5677177 -0.4677177 -0.36771772 -0.267717722
#> 6  06 -0.5518946 -0.4518946 -0.35189458 -0.251894577
#> 7  07 -0.5440045 -0.4440045 -0.34400446 -0.244004456
#> 8  08 -0.2980174 -0.1980174 -0.09801741  0.001982594
#> 9  09 -0.8638542 -0.7638542 -0.66385418 -0.563854183
#> 10 10 -0.8481297 -0.7481297 -0.64812968 -0.548129680

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.63358631 -0.63358631 -0.53358631  -0.3335863   0.2664137
#> 2  02 -0.40772461 -0.40772461 -0.30772461  -0.1077246   0.4922754
#> 3  03  0.15312615  0.15312615  0.25312615   0.4531262   1.0531262
#> 4  04 -2.63547346 -2.63547346 -2.53547346  -2.3354735  -1.7354735
#> 5  05 -0.14063558 -0.14063558 -0.04063558   0.1593644   0.7593644
#> 6  06 -0.03519655 -0.03519655  0.06480345   0.2648035   0.8648035
#> 7  07 -0.58824533 -0.58824533 -0.48824533  -0.2882453   0.3117547
#> 8  08 -0.17379179 -0.17379179 -0.07379179   0.1262082   0.7262082
#> 9  09 -1.23411881 -1.23411881 -1.13411881  -0.9341188  -0.3341188
#> 10 10  1.77395642  1.77395642  1.87395642   2.0739564   2.6739564