Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -1.1645945 -1.1645945 -1.06459447
#> 2  02 -0.7212265 -0.7212265 -0.62122655
#> 3  03 -0.7622792 -0.7622792 -0.66227923
#> 4  04 -0.7056299 -0.7056299 -0.60562988
#> 5  05  0.8574258  0.8574258  0.95742581
#> 6  06  1.4521057  1.4521057  1.55210571
#> 7  07 -1.2095325 -1.2095325 -1.10953246
#> 8  08 -0.4047636 -0.4047636 -0.30476359
#> 9  09 -0.1318111 -0.1318111 -0.03181105
#> 10 10  0.9546523  0.9546523  1.05465232

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.06435503  0.06435503  0.1643550
#> 2  02  0.39260968  0.39260968  0.4926097
#> 3  03  0.04606874  0.04606874  0.1460687
#> 4  04  0.02560201  0.02560201  0.1256020
#> 5  05 -1.28251344 -1.28251344 -1.1825134
#> 6  06 -0.25748489 -0.25748489 -0.1574849
#> 7  07  0.14491594  0.14491594  0.2449159
#> 8  08 -0.84323678 -0.84323678 -0.7432368
#> 9  09  0.51997570  0.51997570  0.6199757
#> 10 10 -0.39299753 -0.39299753 -0.2929975

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -0.37163892 -0.27163892 -0.1716389 -0.07163892
#> 2  02  1.02253858  1.12253858  1.2225386  1.32253858
#> 3  03  0.82827623  0.92827623  1.0282762  1.12827623
#> 4  04 -1.07000469 -0.97000469 -0.8700047 -0.77000469
#> 5  05  0.38122817  0.48122817  0.5812282  0.68122817
#> 6  06 -0.07785332  0.02214668  0.1221467  0.22214668
#> 7  07 -0.03511683  0.06488317  0.1648832  0.26488317
#> 8  08  0.01661751  0.11661751  0.2166175  0.31661751
#> 9  09  0.86913324  0.96913324  1.0691332  1.16913324
#> 10 10  1.66452307  1.76452307  1.8645231  1.96452307

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.94085079 -0.94085079 -0.84085079 -0.64085079 -0.04085079
#> 2  02 -0.31959635 -0.31959635 -0.21959635 -0.01959635  0.58040365
#> 3  03  0.06960734  0.06960734  0.16960734  0.36960734  0.96960734
#> 4  04 -0.01029721 -0.01029721  0.08970279  0.28970279  0.88970279
#> 5  05  0.06126290  0.06126290  0.16126290  0.36126290  0.96126290
#> 6  06  1.46437073  1.46437073  1.56437073  1.76437073  2.36437073
#> 7  07 -0.36368957 -0.36368957 -0.26368957 -0.06368957  0.53631043
#> 8  08 -0.45186463 -0.45186463 -0.35186463 -0.15186463  0.44813537
#> 9  09 -0.83851624 -0.83851624 -0.73851624 -0.53851624  0.06148376
#> 10 10  0.01221895  0.01221895  0.11221895  0.31221895  0.91221895