Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.003432850  0.003432850  0.10343285
#> 2  02 -0.531526440 -0.531526440 -0.43152644
#> 3  03 -0.006736682 -0.006736682  0.09326332
#> 4  04  0.245248295  0.245248295  0.34524829
#> 5  05 -1.296230435 -1.296230435 -1.19623043
#> 6  06  0.024198085  0.024198085  0.12419808
#> 7  07  0.063354718  0.063354718  0.16335472
#> 8  08  0.054850894  0.054850894  0.15485089
#> 9  09  0.877336895  0.877336895  0.97733689
#> 10 10  1.821934320  1.821934320  1.92193432

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -3.06016103 -3.06016103 -2.96016103
#> 2  02 -0.03495708 -0.03495708  0.06504292
#> 3  03  0.93424676  0.93424676  1.03424676
#> 4  04  1.44686884  1.44686884  1.54686884
#> 5  05  0.61384377  0.61384377  0.71384377
#> 6  06 -0.82056115 -0.82056115 -0.72056115
#> 7  07 -0.03793500 -0.03793500  0.06206500
#> 8  08 -0.06945283 -0.06945283  0.03054717
#> 9  09 -1.41349947 -1.41349947 -1.31349947
#> 10 10 -0.74720908 -0.74720908 -0.64720908

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -0.895531088 -0.79553109 -0.69553109 -0.5955311
#> 2  02  0.879822210  0.97982221  1.07982221  1.1798222
#> 3  03  0.074050459  0.17405046  0.27405046  0.3740505
#> 4  04  0.328237847  0.42823785  0.52823785  0.6282378
#> 5  05 -0.880042343 -0.78004234 -0.68004234 -0.5800423
#> 6  06 -0.152385270 -0.05238527  0.04761473  0.1476147
#> 7  07 -0.005034916  0.09496508  0.19496508  0.2949651
#> 8  08 -0.034149428  0.06585057  0.16585057  0.2658506
#> 9  09 -1.094544314 -0.99454431 -0.89454431 -0.7945443
#> 10 10 -0.137924787 -0.03792479  0.06207521  0.1620752

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.6369548  -1.6369548 -1.53695479  -1.3369548 -0.73695479
#> 2  02  0.4561738   0.4561738  0.55617376   0.7561738  1.35617376
#> 3  03  0.1915988   0.1915988  0.29159879   0.4915988  1.09159879
#> 4  04 -1.1389158  -1.1389158 -1.03891575  -0.8389158 -0.23891575
#> 5  05 -0.5334536  -0.5334536 -0.43345365  -0.2334536  0.36654635
#> 6  06 -0.8515972  -0.8515972 -0.75159724  -0.5515972  0.04840276
#> 7  07 -1.6980841  -1.6980841 -1.59808406  -1.3980841 -0.79808406
#> 8  08 -1.1337721  -1.1337721 -1.03377209  -0.8337721 -0.23377209
#> 9  09 -0.9955506  -0.9955506 -0.89555061  -0.6955506 -0.09555061
#> 10 10 -0.1946068  -0.1946068 -0.09460682   0.1053932  0.70539318