Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8568265 -0.8568265 -0.7568265
#> 2  02 -1.1771872 -1.1771872 -1.0771872
#> 3  03 -0.7613497 -0.7613497 -0.6613497
#> 4  04  0.8940005  0.8940005  0.9940005
#> 5  05 -1.9858972 -1.9858972 -1.8858972
#> 6  06  0.3748092  0.3748092  0.4748092
#> 7  07  1.4584276  1.4584276  1.5584276
#> 8  08  0.1395187  0.1395187  0.2395187
#> 9  09 -1.0544003 -1.0544003 -0.9544003
#> 10 10  1.3536283  1.3536283  1.4536283

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.57171221  0.57171221  0.67171221
#> 2  02  2.51650491  2.51650491  2.61650491
#> 3  03  0.03653865  0.03653865  0.13653865
#> 4  04 -0.16440271 -0.16440271 -0.06440271
#> 5  05  0.85711326  0.85711326  0.95711326
#> 6  06 -0.14862816 -0.14862816 -0.04862816
#> 7  07 -0.02343243 -0.02343243  0.07656757
#> 8  08  0.26069626  0.26069626  0.36069626
#> 9  09 -1.64804673 -1.64804673 -1.54804673
#> 10 10  1.95261088  1.95261088  2.05261088

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.6124502  0.7124502  0.8124502  0.91245021
#> 2  02 -0.4774671 -0.3774671 -0.2774671 -0.17746709
#> 3  03 -0.3284689 -0.2284689 -0.1284689 -0.02846894
#> 4  04  1.2562218  1.3562218  1.4562218  1.55622176
#> 5  05 -1.3364894 -1.2364894 -1.1364894 -1.03648941
#> 6  06  0.1822550  0.2822550  0.3822550  0.48225503
#> 7  07 -0.4501601 -0.3501601 -0.2501601 -0.15016008
#> 8  08 -1.9286194 -1.8286194 -1.7286194 -1.62861942
#> 9  09 -1.1440904 -1.0440904 -0.9440904 -0.84409040
#> 10 10  0.2037103  0.3037103  0.4037103  0.50371028

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.7571752  -1.7571752  -1.6571752  -1.4571752  -0.8571752
#> 2  02  0.3045504   0.3045504   0.4045504   0.6045504   1.2045504
#> 3  03 -1.1610724  -1.1610724  -1.0610724  -0.8610724  -0.2610724
#> 4  04  1.5727597   1.5727597   1.6727597   1.8727597   2.4727597
#> 5  05 -1.5459047  -1.5459047  -1.4459047  -1.2459047  -0.6459047
#> 6  06  0.1693634   0.1693634   0.2693634   0.4693634   1.0693634
#> 7  07  1.9702454   1.9702454   2.0702454   2.2702454   2.8702454
#> 8  08  0.1958894   0.1958894   0.2958894   0.4958894   1.0958894
#> 9  09 -1.4623327  -1.4623327  -1.3623327  -1.1623327  -0.5623327
#> 10 10  0.5857639   0.5857639   0.6857639   0.8857639   1.4857639