Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.02839468 -0.02839468  0.07160532
#> 2  02  0.07138230  0.07138230  0.17138230
#> 3  03 -1.22956795 -1.22956795 -1.12956795
#> 4  04  1.42871627  1.42871627  1.52871627
#> 5  05  1.60627696  1.60627696  1.70627696
#> 6  06  0.89937356  0.89937356  0.99937356
#> 7  07  0.86745549  0.86745549  0.96745549
#> 8  08 -0.56278998 -0.56278998 -0.46278998
#> 9  09 -0.18353310 -0.18353310 -0.08353310
#> 10 10 -0.76781773 -0.76781773 -0.66781773

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.81228961  0.81228961  0.91228961
#> 2  02  1.33872029  1.33872029  1.43872029
#> 3  03 -0.65318649 -0.65318649 -0.55318649
#> 4  04 -0.33257861 -0.33257861 -0.23257861
#> 5  05 -1.77940160 -1.77940160 -1.67940160
#> 6  06  0.87045603  0.87045603  0.97045603
#> 7  07 -0.70729449 -0.70729449 -0.60729449
#> 8  08 -0.04486022 -0.04486022  0.05513978
#> 9  09  0.58861512  0.58861512  0.68861512
#> 10 10 -0.39347907 -0.39347907 -0.29347907

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.6193851 -0.51938506 -0.41938506 -0.31938506
#> 2  02 -1.9758266 -1.87582662 -1.77582662 -1.67582662
#> 3  03 -1.4678594 -1.36785942 -1.26785942 -1.16785942
#> 4  04 -0.1621046 -0.06210464  0.03789536  0.13789536
#> 5  05  0.1431022  0.24310218  0.34310218  0.44310218
#> 6  06  2.8558485  2.95584852  3.05584852  3.15584852
#> 7  07  0.0655022  0.16550220  0.26550220  0.36550220
#> 8  08 -0.3659438 -0.26594375 -0.16594375 -0.06594375
#> 9  09  2.1297196  2.22971960  2.32971960  2.42971960
#> 10 10 -0.4298576 -0.32985762 -0.22985762 -0.12985762

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.06802719  0.06802719   0.1680272   0.3680272  0.96802719
#> 2  02  1.95264482  1.95264482   2.0526448   2.2526448  2.85264482
#> 3  03 -1.22752414 -1.22752414  -1.1275241  -0.9275241 -0.32752414
#> 4  04 -1.03543963 -1.03543963  -0.9354396  -0.7354396 -0.13543963
#> 5  05  0.09246850  0.09246850   0.1924685   0.3924685  0.99246850
#> 6  06 -0.96836023 -0.96836023  -0.8683602  -0.6683602 -0.06836023
#> 7  07  0.94390864  0.94390864   1.0439086   1.2439086  1.84390864
#> 8  08 -2.18079304 -2.18079304  -2.0807930  -1.8807930 -1.28079304
#> 9  09  0.81980919  0.81980919   0.9198092   1.1198092  1.71980919
#> 10 10  0.69927734  0.69927734   0.7992773   0.9992773  1.59927734