Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.3696061  0.3696061  0.46960605
#> 2  02 -1.5745564 -1.5745564 -1.47455641
#> 3  03 -0.1117448 -0.1117448 -0.01174484
#> 4  04 -0.7101108 -0.7101108 -0.61011084
#> 5  05  0.2614296  0.2614296  0.36142955
#> 6  06  1.0932532  1.0932532  1.19325322
#> 7  07 -2.4171207 -2.4171207 -2.31712073
#> 8  08 -0.9860272 -0.9860272 -0.88602722
#> 9  09  0.3584498  0.3584498  0.45844979
#> 10 10  2.0095483  2.0095483  2.10954831

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.7358904 -0.7358904 -0.6358904
#> 2  02  0.9721396  0.9721396  1.0721396
#> 3  03 -0.8745392 -0.8745392 -0.7745392
#> 4  04  0.1452143  0.1452143  0.2452143
#> 5  05  0.8703540  0.8703540  0.9703540
#> 6  06  1.7433273  1.7433273  1.8433273
#> 7  07  1.0435941  1.0435941  1.1435941
#> 8  08  0.9257610  0.9257610  1.0257610
#> 9  09  0.8156586  0.8156586  0.9156586
#> 10 10 -0.4262150 -0.4262150 -0.3262150

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.4280672 -0.3280672 -0.2280672 -0.1280672
#> 2  02  0.5164568  0.6164568  0.7164568  0.8164568
#> 3  03 -0.6498572 -0.5498572 -0.4498572 -0.3498572
#> 4  04 -1.9996418 -1.8996418 -1.7996418 -1.6996418
#> 5  05  0.4548929  0.5548929  0.6548929  0.7548929
#> 6  06  1.3768492  1.4768492  1.5768492  1.6768492
#> 7  07  0.9853680  1.0853680  1.1853680  1.2853680
#> 8  08 -1.2544897 -1.1544897 -1.0544897 -0.9544897
#> 9  09  0.9259040  1.0259040  1.1259040  1.2259040
#> 10 10  1.4548052  1.5548052  1.6548052  1.7548052

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.4464591   0.4464591   0.5464591  0.74645907   1.3464591
#> 2  02 -0.6597998  -0.6597998  -0.5597998 -0.35979983   0.2402002
#> 3  03  0.8832741   0.8832741   0.9832741  1.18327407   1.7832741
#> 4  04 -1.6175634  -1.6175634  -1.5175634 -1.31756340  -0.7175634
#> 5  05 -0.3306139  -0.3306139  -0.2306139 -0.03061387   0.5693861
#> 6  06 -1.2450665  -1.2450665  -1.1450665 -0.94506651  -0.3450665
#> 7  07  0.7082800   0.7082800   0.8082800  1.00828002   1.6082800
#> 8  08  1.3303380   1.3303380   1.4303380  1.63033804   2.2303380
#> 9  09  2.4483697   2.4483697   2.5483697  2.74836973   3.3483697
#> 10 10 -2.9434504  -2.9434504  -2.8434504 -2.64345035  -2.0434504