Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.9033587 -0.9033587 -0.80335867
#> 2  02 -0.1518637 -0.1518637 -0.05186372
#> 3  03 -0.2322101 -0.2322101 -0.13221013
#> 4  04  0.4649121  0.4649121  0.56491212
#> 5  05  1.3588984  1.3588984  1.45889842
#> 6  06 -1.0053172 -1.0053172 -0.90531723
#> 7  07  1.1685553  1.1685553  1.26855529
#> 8  08  2.7700840  2.7700840  2.87008404
#> 9  09  1.0551442  1.0551442  1.15514416
#> 10 10 -1.4118428 -1.4118428 -1.31184283

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.48466634 -0.48466634 -0.3846663
#> 2  02  0.22933337  0.22933337  0.3293334
#> 3  03  0.53674492  0.53674492  0.6367449
#> 4  04 -0.40839169 -0.40839169 -0.3083917
#> 5  05 -0.48399183 -0.48399183 -0.3839918
#> 6  06  0.09726679  0.09726679  0.1972668
#> 7  07  0.36490615  0.36490615  0.4649061
#> 8  08  0.47365011  0.47365011  0.5736501
#> 9  09 -0.63866931 -0.63866931 -0.5386693
#> 10 10  0.01168704  0.01168704  0.1116870

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.12315916  0.2231592  0.32315916  0.42315916
#> 2  02  1.85042203  1.9504220  2.05042203  2.15042203
#> 3  03 -1.95081060 -1.8508106 -1.75081060 -1.65081060
#> 4  04 -0.82700134 -0.7270013 -0.62700134 -0.52700134
#> 5  05  0.36809105  0.4680910  0.56809105  0.66809105
#> 6  06  0.90143163  1.0014316  1.10143163  1.20143163
#> 7  07 -0.97177859 -0.8717786 -0.77177859 -0.67177859
#> 8  08  0.02692345  0.1269235  0.22692345  0.32692345
#> 9  09  1.82115318  1.9211532  2.02115318  2.12115318
#> 10 10 -0.26939264 -0.1693926 -0.06939264  0.03060736

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.2203956  -1.2203956  -1.1203956 -0.92039558  -0.3203956
#> 2  02  0.2946366   0.2946366   0.3946366  0.59463658   1.1946366
#> 3  03 -2.8134176  -2.8134176  -2.7134176 -2.51341758  -1.9134176
#> 4  04 -2.0431011  -2.0431011  -1.9431011 -1.74310113  -1.1431011
#> 5  05 -0.7508373  -0.7508373  -0.6508373 -0.45083731   0.1491627
#> 6  06  0.1089948   0.1089948   0.2089948  0.40899483   1.0089948
#> 7  07 -0.4625583  -0.4625583  -0.3625583 -0.16255832   0.4374417
#> 8  08  0.4569326   0.4569326   0.5569326  0.75693263   1.3569326
#> 9  09 -0.2112498  -0.2112498  -0.1112498  0.08875025   0.6887502
#> 10 10 -1.1132986  -1.1132986  -1.0132986 -0.81329858  -0.2132986