Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01 -0.381742863 -0.381742863 -0.2817429
#> 2  02 -0.312995280 -0.312995280 -0.2129953
#> 3  03 -1.065588546 -1.065588546 -0.9655885
#> 4  04  1.025307997  1.025307997  1.1253080
#> 5  05 -0.665623324 -0.665623324 -0.5656233
#> 6  06  0.669213277  0.669213277  0.7692133
#> 7  07  0.082049751  0.082049751  0.1820498
#> 8  08 -0.595474944 -0.595474944 -0.4954749
#> 9  09 -1.054671649 -1.054671649 -0.9546716
#> 10 10  0.005417731  0.005417731  0.1054177

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.43339961 -1.43339961 -1.33339961
#> 2  02 -0.81191122 -0.81191122 -0.71191122
#> 3  03  0.96406462  0.96406462  1.06406462
#> 4  04  0.08613180  0.08613180  0.18613180
#> 5  05 -0.05494372 -0.05494372  0.04505628
#> 6  06  0.10999263  0.10999263  0.20999263
#> 7  07 -1.56835798 -1.56835798 -1.46835798
#> 8  08  0.93370176  0.93370176  1.03370176
#> 9  09  1.28338198  1.28338198  1.38338198
#> 10 10  0.57703427  0.57703427  0.67703427

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.84420061  0.94420061  1.0442006  1.14420061
#> 2  02 -1.33053162 -1.23053162 -1.1305316 -1.03053162
#> 3  03 -1.27003100 -1.17003100 -1.0700310 -0.97003100
#> 4  04 -1.49501036 -1.39501036 -1.2950104 -1.19501036
#> 5  05 -0.37023923 -0.27023923 -0.1702392 -0.07023923
#> 6  06 -0.08277338  0.01722662  0.1172266  0.21722662
#> 7  07  0.23625438  0.33625438  0.4362544  0.53625438
#> 8  08  0.95419099  1.05419099  1.1541910  1.25419099
#> 9  09  1.13676431  1.23676431  1.3367643  1.43676431
#> 10 10 -0.66883419 -0.56883419 -0.4688342 -0.36883419

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.85096219  1.85096219  1.95096219   2.1509622   2.7509622
#> 2  02  0.09255275  0.09255275  0.19255275   0.3925528   0.9925528
#> 3  03 -0.19295262 -0.19295262 -0.09295262   0.1070474   0.7070474
#> 4  04  0.70562772  0.70562772  0.80562772   1.0056277   1.6056277
#> 5  05  1.41631892  1.41631892  1.51631892   1.7163189   2.3163189
#> 6  06 -0.45966735 -0.45966735 -0.35966735  -0.1596673   0.4403327
#> 7  07 -2.12900408 -2.12900408 -2.02900408  -1.8290041  -1.2290041
#> 8  08  0.23637761  0.23637761  0.33637761   0.5363776   1.1363776
#> 9  09  0.46974060  0.46974060  0.56974060   0.7697406   1.3697406
#> 10 10 -1.45904315 -1.45904315 -1.35904315  -1.1590431  -0.5590431