Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.4975406 -0.4975406 -0.3975406
#> 2  02  0.6597512  0.6597512  0.7597512
#> 3  03  0.7557717  0.7557717  0.8557717
#> 4  04 -0.2531278 -0.2531278 -0.1531278
#> 5  05 -0.3034932 -0.3034932 -0.2034932
#> 6  06 -0.3253672 -0.3253672 -0.2253672
#> 7  07 -0.8665782 -0.8665782 -0.7665782
#> 8  08  0.6133259  0.6133259  0.7133259
#> 9  09 -0.3842537 -0.3842537 -0.2842537
#> 10 10 -0.0608925 -0.0608925  0.0391075

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01 -0.853341273 -0.853341273 -0.75334127
#> 2  02  0.003019723  0.003019723  0.10301972
#> 3  03 -0.390250175 -0.390250175 -0.29025017
#> 4  04 -0.464322873 -0.464322873 -0.36432287
#> 5  05 -0.020805653 -0.020805653  0.07919435
#> 6  06  0.857203072  0.857203072  0.95720307
#> 7  07  0.276868907  0.276868907  0.37686891
#> 8  08 -0.639432139 -0.639432139 -0.53943214
#> 9  09  0.241608168  0.241608168  0.34160817
#> 10 10  0.667163392  0.667163392  0.76716339

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -1.73302936 -1.63302936 -1.53302936 -1.4330294
#> 2  02 -0.04615633  0.05384367  0.15384367  0.2538437
#> 3  03 -1.82826209 -1.72826209 -1.62826209 -1.5282621
#> 4  04 -2.00697616 -1.90697616 -1.80697616 -1.7069762
#> 5  05  0.75638800  0.85638800  0.95638800  1.0563880
#> 6  06  2.42530702  2.52530702  2.62530702  2.7253070
#> 7  07 -0.73986962 -0.63986962 -0.53986962 -0.4398696
#> 8  08 -0.16507841 -0.06507841  0.03492159  0.1349216
#> 9  09 -1.48397997 -1.38397997 -1.28397997 -1.1839800
#> 10 10  0.60253350  0.70253350  0.80253350  0.9025335

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.815018416 -0.815018416  -0.7150184  -0.5150184  0.08498158
#> 2  02 -2.237006225 -2.237006225  -2.1370062  -1.9370062 -1.33700622
#> 3  03 -1.056504230 -1.056504230  -0.9565042  -0.7565042 -0.15650423
#> 4  04  1.109552337  1.109552337   1.2095523   1.4095523  2.00955234
#> 5  05  1.604540598  1.604540598   1.7045406   1.9045406  2.50454060
#> 6  06  0.419185765  0.419185765   0.5191858   0.7191858  1.31918577
#> 7  07  0.476296391  0.476296391   0.5762964   0.7762964  1.37629639
#> 8  08  0.332633176  0.332633176   0.4326332   0.6326332  1.23263318
#> 9  09  0.006747015  0.006747015   0.1067470   0.3067470  0.90674701
#> 10 10  0.184644684  0.184644684   0.2846447   0.4846447  1.08464468