Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.41599883  1.41599883  1.5159988
#> 2  02 -0.73429737 -0.73429737 -0.6342974
#> 3  03 -0.25857748 -0.25857748 -0.1585775
#> 4  04  0.23864223  0.23864223  0.3386422
#> 5  05  1.02989777  1.02989777  1.1298978
#> 6  06  0.07483447  0.07483447  0.1748345
#> 7  07  0.56710928  0.56710928  0.6671093
#> 8  08 -1.78239874 -1.78239874 -1.6823987
#> 9  09  0.19311756  0.19311756  0.2931176
#> 10 10 -1.38132409 -1.38132409 -1.2813241

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.7311794  0.7311794  0.8311794
#> 2  02 -0.2349013 -0.2349013 -0.1349013
#> 3  03 -0.3988379 -0.3988379 -0.2988379
#> 4  04 -1.9589467 -1.9589467 -1.8589467
#> 5  05  1.3776885  1.3776885  1.4776885
#> 6  06 -0.8218052 -0.8218052 -0.7218052
#> 7  07 -0.8872740 -0.8872740 -0.7872740
#> 8  08 -0.8696158 -0.8696158 -0.7696158
#> 9  09 -0.3855990 -0.3855990 -0.2855990
#> 10 10  2.1296816  2.1296816  2.2296816

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01 -0.2943390 -0.1943390 -0.09433904  0.005660964
#> 2  02 -0.6679157 -0.5679157 -0.46791571 -0.367915714
#> 3  03  1.4663956  1.5663956  1.66639564  1.766395644
#> 4  04 -1.5420816 -1.4420816 -1.34208162 -1.242081622
#> 5  05 -0.5776560 -0.4776560 -0.37765595 -0.277655953
#> 6  06 -1.3729916 -1.2729916 -1.17299156 -1.072991557
#> 7  07  0.2347350  0.3347350  0.43473503  0.534735028
#> 8  08 -0.6763592 -0.5763592 -0.47635918 -0.376359184
#> 9  09  0.2871596  0.3871596  0.48715962  0.587159619
#> 10 10  1.3665780  1.4665780  1.56657800  1.666578002

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.7742135  -0.7742135  -0.6742135  -0.4742135   0.1257865
#> 2  02 -0.7986556  -0.7986556  -0.6986556  -0.4986556   0.1013444
#> 3  03  0.2520002   0.2520002   0.3520002   0.5520002   1.1520002
#> 4  04 -1.3268960  -1.3268960  -1.2268960  -1.0268960  -0.4268960
#> 5  05 -0.6043604  -0.6043604  -0.5043604  -0.3043604   0.2956396
#> 6  06  0.1263701   0.1263701   0.2263701   0.4263701   1.0263701
#> 7  07  0.2944457   0.2944457   0.3944457   0.5944457   1.1944457
#> 8  08  0.8881111   0.8881111   0.9881111   1.1881111   1.7881111
#> 9  09  1.9338338   1.9338338   2.0338338   2.2338338   2.8338338
#> 10 10  1.1262701   1.1262701   1.2262701   1.4262701   2.0262701