Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01 -0.430651593 -0.430651593 -0.33065159
#> 2  02 -0.004581141 -0.004581141  0.09541886
#> 3  03 -0.311730974 -0.311730974 -0.21173097
#> 4  04 -1.539966213 -1.539966213 -1.43996621
#> 5  05 -0.894300652 -0.894300652 -0.79430065
#> 6  06  0.064211594  0.064211594  0.16421159
#> 7  07  0.678982815  0.678982815  0.77898282
#> 8  08  0.539471537  0.539471537  0.63947154
#> 9  09  0.393041308  0.393041308  0.49304131
#> 10 10 -1.271192729 -1.271192729 -1.17119273

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.78432375  0.78432375  0.88432375
#> 2  02  0.12766684  0.12766684  0.22766684
#> 3  03  1.51630994  1.51630994  1.61630994
#> 4  04 -0.12597731 -0.12597731 -0.02597731
#> 5  05  0.13576511  0.13576511  0.23576511
#> 6  06 -0.33405983 -0.33405983 -0.23405983
#> 7  07  0.10633102  0.10633102  0.20633102
#> 8  08  0.31680978  0.31680978  0.41680978
#> 9  09 -0.06983160 -0.06983160  0.03016840
#> 10 10  0.06847263  0.06847263  0.16847263

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  0.39810498  0.498104984  0.59810498  0.6981050
#> 2  02 -1.23922495 -1.139224951 -1.03922495 -0.9392250
#> 3  03  1.07126822  1.171268225  1.27126822  1.3712682
#> 4  04 -0.43718477 -0.337184773 -0.23718477 -0.1371848
#> 5  05  0.44779729  0.547797292  0.64779729  0.7477973
#> 6  06 -0.09545164  0.004548361  0.10454836  0.2045484
#> 7  07 -1.32264866 -1.222648660 -1.12264866 -1.0226487
#> 8  08 -0.14220693 -0.042206929  0.05779307  0.1577931
#> 9  09  0.06057090  0.160570903  0.26057090  0.3605709
#> 10 10 -0.60414804 -0.504148035 -0.40414804 -0.3041480

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.12264850  0.12264850  0.22264850   0.4226485   1.0226485
#> 2  02 -1.01644476 -1.01644476 -0.91644476  -0.7164448  -0.1164448
#> 3  03 -0.67281666 -0.67281666 -0.57281666  -0.3728167   0.2271833
#> 4  04  1.06194400  1.06194400  1.16194400   1.3619440   1.9619440
#> 5  05  0.31590117  0.31590117  0.41590117   0.6159012   1.2159012
#> 6  06 -0.41768627 -0.41768627 -0.31768627  -0.1176863   0.4823137
#> 7  07  1.28827902  1.28827902  1.38827902   1.5882790   2.1882790
#> 8  08 -0.08587859 -0.08587859  0.01412141   0.2141214   0.8141214
#> 9  09 -1.13177612 -1.13177612 -1.03177612  -0.8317761  -0.2317761
#> 10 10  1.75943144  1.75943144  1.85943144   2.0594314   2.6594314