Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.012541992  0.012541992  0.11254199
#> 2  02 -0.009579674 -0.009579674  0.09042033
#> 3  03 -0.254369090 -0.254369090 -0.15436909
#> 4  04  0.417992384  0.417992384  0.51799238
#> 5  05  1.233945680  1.233945680  1.33394568
#> 6  06  0.838508924  0.838508924  0.93850892
#> 7  07  0.782816685  0.782816685  0.88281668
#> 8  08  0.531143446  0.531143446  0.63114345
#> 9  09 -1.527439132 -1.527439132 -1.42743913
#> 10 10 -0.415485216 -0.415485216 -0.31548522

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.1481509  1.1481509  1.2481509
#> 2  02 -1.5820785 -1.5820785 -1.4820785
#> 3  03 -0.2636421 -0.2636421 -0.1636421
#> 4  04 -0.4552856 -0.4552856 -0.3552856
#> 5  05 -0.7734189 -0.7734189 -0.6734189
#> 6  06 -0.8781132 -0.8781132 -0.7781132
#> 7  07  0.5598143  0.5598143  0.6598143
#> 8  08  0.3139859  0.3139859  0.4139859
#> 9  09  1.7396988  1.7396988  1.8396988
#> 10 10 -0.4411405 -0.4411405 -0.3411405

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.121364109  0.2213641  0.3213641  0.4213641
#> 2  02  0.809692518  0.9096925  1.0096925  1.1096925
#> 3  03  0.005675768  0.1056758  0.2056758  0.3056758
#> 4  04  0.118192222  0.2181922  0.3181922  0.4181922
#> 5  05  0.257321943  0.3573219  0.4573219  0.5573219
#> 6  06  0.024946175  0.1249462  0.2249462  0.3249462
#> 7  07 -0.140554996 -0.0405550  0.0594450  0.1594450
#> 8  08 -0.511999453 -0.4119995 -0.3119995 -0.2119995
#> 9  09  0.026493158  0.1264932  0.2264932  0.3264932
#> 10 10 -0.692515474 -0.5925155 -0.4925155 -0.3925155

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.7014316  -0.7014316  -0.6014316 -0.40143156   0.1985684
#> 2  02  0.7104362   0.7104362   0.8104362  1.01043623   1.6104362
#> 3  03  0.7715509   0.7715509   0.8715509  1.07155093   1.6715509
#> 4  04 -0.2150031  -0.2150031  -0.1150031  0.08499691   0.6849969
#> 5  05  0.5435656   0.5435656   0.6435656  0.84356561   1.4435656
#> 6  06  0.5398010   0.5398010   0.6398010  0.83980097   1.4398010
#> 7  07 -0.4058227  -0.4058227  -0.3058227 -0.10582270   0.4941773
#> 8  08 -1.3077390  -1.3077390  -1.2077390 -1.00773902  -0.4077390
#> 9  09  0.1366465   0.1366465   0.2366465  0.43664652   1.0366465
#> 10 10  1.2535985   1.2535985   1.3535985  1.55359846   2.1535985