Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.4969577 -1.4969577 -1.3969577
#> 2  02 -0.6843813 -0.6843813 -0.5843813
#> 3  03  0.1616787  0.1616787  0.2616787
#> 4  04 -1.3439615 -1.3439615 -1.2439615
#> 5  05 -2.9614967 -2.9614967 -2.8614967
#> 6  06 -1.5521282 -1.5521282 -1.4521282
#> 7  07 -1.0440830 -1.0440830 -0.9440830
#> 8  08  1.4549785  1.4549785  1.5549785
#> 9  09 -0.4242272 -0.4242272 -0.3242272
#> 10 10 -1.4739535 -1.4739535 -1.3739535

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.2961328  1.2961328  1.3961328
#> 2  02 -0.3162319 -0.3162319 -0.2162319
#> 3  03  0.1875736  0.1875736  0.2875736
#> 4  04 -0.5915097 -0.5915097 -0.4915097
#> 5  05  0.1857471  0.1857471  0.2857471
#> 6  06  1.2780915  1.2780915  1.3780915
#> 7  07  1.7425199  1.7425199  1.8425199
#> 8  08  0.8431661  0.8431661  0.9431661
#> 9  09  0.8543741  0.8543741  0.9543741
#> 10 10 -0.2722043 -0.2722043 -0.1722043

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.20005678  1.30005678  1.40005678  1.50005678
#> 2  02 -1.08845888 -0.98845888 -0.88845888 -0.78845888
#> 3  03  0.07770018  0.17770018  0.27770018  0.37770018
#> 4  04 -1.26599700 -1.16599700 -1.06599700 -0.96599700
#> 5  05 -0.28366397 -0.18366397 -0.08366397  0.01633603
#> 6  06  0.76719984  0.86719984  0.96719984  1.06719984
#> 7  07  0.36239194  0.46239194  0.56239194  0.66239194
#> 8  08  0.70332337  0.80332337  0.90332337  1.00332337
#> 9  09  1.40303862  1.50303862  1.60303862  1.70303862
#> 10 10 -0.11460193 -0.01460193  0.08539807  0.18539807

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.1942604   0.1942604  0.29426036   0.4942604   1.0942604
#> 2  02 -1.5223893  -1.5223893 -1.42238925  -1.2223893  -0.6223893
#> 3  03 -0.1271262  -0.1271262 -0.02712622   0.1728738   0.7728738
#> 4  04  1.0671017   1.0671017  1.16710171   1.3671017   1.9671017
#> 5  05 -0.1224764  -0.1224764 -0.02247635   0.1775236   0.7775236
#> 6  06  1.5084280   1.5084280  1.60842800   1.8084280   2.4084280
#> 7  07  0.7516943   0.7516943  0.85169428   1.0516943   1.6516943
#> 8  08 -0.4805992  -0.4805992 -0.38059917  -0.1805992   0.4194008
#> 9  09 -0.4764539  -0.4764539 -0.37645392  -0.1764539   0.4235461
#> 10 10  1.5710253   1.5710253  1.67102527   1.8710253   2.4710253