Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01 -0.905362951 -0.905362951 -0.80536295
#> 2  02  0.004135943  0.004135943  0.10413594
#> 3  03  1.192579788  1.192579788  1.29257979
#> 4  04 -0.179232760 -0.179232760 -0.07923276
#> 5  05 -0.463613061 -0.463613061 -0.36361306
#> 6  06 -1.099873062 -1.099873062 -0.99987306
#> 7  07 -0.364516713 -0.364516713 -0.26451671
#> 8  08 -0.201019514 -0.201019514 -0.10101951
#> 9  09  0.080623803  0.080623803  0.18062380
#> 10 10 -1.450342713 -1.450342713 -1.35034271

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.07764164  1.07764164  1.17764164
#> 2  02  1.15834747  1.15834747  1.25834747
#> 3  03 -1.83021736 -1.83021736 -1.73021736
#> 4  04 -1.22579270 -1.22579270 -1.12579270
#> 5  05  0.05671029  0.05671029  0.15671029
#> 6  06  1.19026205  1.19026205  1.29026205
#> 7  07  0.91405538  0.91405538  1.01405538
#> 8  08 -0.01653488 -0.01653488  0.08346512
#> 9  09 -0.45161389 -0.45161389 -0.35161389
#> 10 10 -0.80562248 -0.80562248 -0.70562248

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.03556887  0.06443113  0.1644311  0.2644311
#> 2  02  2.15018159  2.25018159  2.3501816  2.4501816
#> 3  03  0.02877329  0.12877329  0.2287733  0.3287733
#> 4  04 -0.96060060 -0.86060060 -0.7606006 -0.6606006
#> 5  05 -1.03526954 -0.93526954 -0.8352695 -0.7352695
#> 6  06  0.25957287  0.35957287  0.4595729  0.5595729
#> 7  07  0.68189986  0.78189986  0.8818999  0.9818999
#> 8  08 -0.82630932 -0.72630932 -0.6263093 -0.5263093
#> 9  09  0.78718942  0.88718942  0.9871894  1.0871894
#> 10 10 -1.73960747 -1.63960747 -1.5396075 -1.4396075

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -2.0099225  -2.0099225  -1.9099225 -1.70992249  -1.1099225
#> 2  02 -1.2066838  -1.2066838  -1.1066838 -0.90668381  -0.3066838
#> 3  03  1.0159130   1.0159130   1.1159130  1.31591299   1.9159130
#> 4  04 -1.2110292  -1.2110292  -1.1110292 -0.91102921  -0.3110292
#> 5  05  1.1321214   1.1321214   1.2321214  1.43212144   2.0321214
#> 6  06 -0.3886406  -0.3886406  -0.2886406 -0.08864058   0.5113594
#> 7  07  0.5875729   0.5875729   0.6875729  0.88757289   1.4875729
#> 8  08 -1.0058339  -1.0058339  -0.9058339 -0.70583388  -0.1058339
#> 9  09  1.1387086   1.1387086   1.2387086  1.43870856   2.0387086
#> 10 10  0.4470225   0.4470225   0.5470225  0.74702253   1.3470225