Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.04779647 -0.04779647  0.05220353
#> 2  02  0.12931210  0.12931210  0.22931210
#> 3  03 -0.45467261 -0.45467261 -0.35467261
#> 4  04  2.36500732  2.36500732  2.46500732
#> 5  05  0.01488576  0.01488576  0.11488576
#> 6  06 -2.01082933 -2.01082933 -1.91082933
#> 7  07  1.10862481  1.10862481  1.20862481
#> 8  08  0.02789393  0.02789393  0.12789393
#> 9  09 -1.00212588 -1.00212588 -0.90212588
#> 10 10 -0.18837200 -0.18837200 -0.08837200

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.56094785 -0.56094785 -0.4609478
#> 2  02  1.26744789  1.26744789  1.3674479
#> 3  03 -0.50001976 -0.50001976 -0.4000198
#> 4  04  1.52289985  1.52289985  1.6228998
#> 5  05  0.05932937  0.05932937  0.1593294
#> 6  06 -2.05230598 -2.05230598 -1.9523060
#> 7  07  1.23168519  1.23168519  1.3316852
#> 8  08 -0.73325602 -0.73325602 -0.6332560
#> 9  09  0.38364920  0.38364920  0.4836492
#> 10 10  0.58449381  0.58449381  0.6844938

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.3530846  0.4530846  0.55308458  0.65308458
#> 2  02 -0.2751954 -0.1751954 -0.07519542  0.02480458
#> 3  03  0.9710588  1.0710588  1.17105877  1.27105877
#> 4  04 -1.1040709 -1.0040709 -0.90407093 -0.80407093
#> 5  05  1.1515031  1.2515031  1.35150313  1.45150313
#> 6  06 -1.2733896 -1.1733896 -1.07338957 -0.97338957
#> 7  07  1.3281907  1.4281907  1.52819072  1.62819072
#> 8  08  0.6699299  0.7699299  0.86992992  0.96992992
#> 9  09 -0.7840031 -0.6840031 -0.58400314 -0.48400314
#> 10 10 -1.6920638 -1.5920638 -1.49206382 -1.39206382

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.8116078  -0.8116078  -0.7116078 -0.511607755  0.08839225
#> 2  02 -1.3954772  -1.3954772  -1.2954772 -1.095477238 -0.49547724
#> 3  03  0.3781142   0.3781142   0.4781142  0.678114164  1.27811416
#> 4  04  0.4914436   0.4914436   0.5914436  0.791443570  1.39144357
#> 5  05  0.4098191   0.4098191   0.5098191  0.709819144  1.30981914
#> 6  06  0.6152090   0.6152090   0.7152090  0.915209026  1.51520903
#> 7  07  0.4397178   0.4397178   0.5397178  0.739717790  1.33971779
#> 8  08 -0.3488815  -0.3488815  -0.2488815 -0.048881494  0.55111851
#> 9  09  0.1100031   0.1100031   0.2100031  0.410003112  1.01000311
#> 10 10 -0.2984950  -0.2984950  -0.1984950  0.001505035  0.60150504