Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.1307895  0.1307895  0.2307895
#> 2  02  0.4205740  0.4205740  0.5205740
#> 3  03 -0.7225790 -0.7225790 -0.6225790
#> 4  04  0.2592750  0.2592750  0.3592750
#> 5  05  1.2885110  1.2885110  1.3885110
#> 6  06  1.0919931  1.0919931  1.1919931
#> 7  07  1.2551836  1.2551836  1.3551836
#> 8  08  0.5848375  0.5848375  0.6848375
#> 9  09 -1.6059605 -1.6059605 -1.5059605
#> 10 10 -0.9663994 -0.9663994 -0.8663994

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.08201672  1.08201672  1.18201672
#> 2  02  0.51916216  0.51916216  0.61916216
#> 3  03  0.31818458  0.31818458  0.41818458
#> 4  04  0.33760774  0.33760774  0.43760774
#> 5  05  0.99678083  0.99678083  1.09678083
#> 6  06 -1.03984513 -1.03984513 -0.93984513
#> 7  07  0.12044322  0.12044322  0.22044322
#> 8  08  0.33766186  0.33766186  0.43766186
#> 9  09 -2.10005898 -2.10005898 -2.00005898
#> 10 10 -0.08793135 -0.08793135  0.01206865

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.76732625  0.8673262  0.9673262  1.0673262
#> 2  02  0.58509036  0.6850904  0.7850904  0.8850904
#> 3  03  0.05458881  0.1545888  0.2545888  0.3545888
#> 4  04  0.07552971  0.1755297  0.2755297  0.3755297
#> 5  05 -0.74374749 -0.6437475 -0.5437475 -0.4437475
#> 6  06  0.67269103  0.7726910  0.8726910  0.9726910
#> 7  07 -1.05669668 -0.9566967 -0.8566967 -0.7566967
#> 8  08  0.73868723  0.8386872  0.9386872  1.0386872
#> 9  09  0.62489422  0.7248942  0.8248942  0.9248942
#> 10 10  0.77408918  0.8740892  0.9740892  1.0740892

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.1416191   0.1416191   0.2416191   0.4416191   1.0416191
#> 2  02  2.5582541   2.5582541   2.6582541   2.8582541   3.4582541
#> 3  03  0.9712705   0.9712705   1.0712705   1.2712705   1.8712705
#> 4  04 -1.2221534  -1.2221534  -1.1221534  -0.9221534  -0.3221534
#> 5  05 -0.8515196  -0.8515196  -0.7515196  -0.5515196   0.0484804
#> 6  06  0.2285608   0.2285608   0.3285608   0.5285608   1.1285608
#> 7  07 -2.2615521  -2.2615521  -2.1615521  -1.9615521  -1.3615521
#> 8  08  0.8529628   0.8529628   0.9529628   1.1529628   1.7529628
#> 9  09  0.1702668   0.1702668   0.2702668   0.4702668   1.0702668
#> 10 10  0.2671295   0.2671295   0.3671295   0.5671295   1.1671295