Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.30638270 -0.30638270 -0.20638270
#> 2  02 -0.15146552 -0.15146552 -0.05146552
#> 3  03  0.21450788  0.21450788  0.31450788
#> 4  04 -0.76610441 -0.76610441 -0.66610441
#> 5  05  0.08279964  0.08279964  0.18279964
#> 6  06 -0.46835660 -0.46835660 -0.36835660
#> 7  07 -0.01252547 -0.01252547  0.08747453
#> 8  08 -0.01238240 -0.01238240  0.08761760
#> 9  09 -0.17179997 -0.17179997 -0.07179997
#> 10 10 -1.79447937 -1.79447937 -1.69447937

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.95359398  1.95359398  2.05359398
#> 2  02  0.07421456  0.07421456  0.17421456
#> 3  03 -0.11682709 -0.11682709 -0.01682709
#> 4  04 -0.51520015 -0.51520015 -0.41520015
#> 5  05 -1.03296426 -1.03296426 -0.93296426
#> 6  06  0.44644858  0.44644858  0.54644858
#> 7  07 -0.29730252 -0.29730252 -0.19730252
#> 8  08  0.49638129  0.49638129  0.59638129
#> 9  09 -0.18594450 -0.18594450 -0.08594450
#> 10 10  0.43596253  0.43596253  0.53596253

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.63828779 -0.53828779 -0.4382878 -0.3382878
#> 2  02 -1.80785996 -1.70785996 -1.6078600 -1.5078600
#> 3  03 -0.06780921  0.03219079  0.1321908  0.2321908
#> 4  04  1.27090659  1.37090659  1.4709066  1.5709066
#> 5  05 -0.81146034 -0.71146034 -0.6114603 -0.5114603
#> 6  06  1.01324763  1.11324763  1.2132476  1.3132476
#> 7  07  1.18148588  1.28148588  1.3814859  1.4814859
#> 8  08  0.13680020  0.23680020  0.3368002  0.4368002
#> 9  09 -1.09552474 -0.99552474 -0.8955247 -0.7955247
#> 10 10  1.08296607  1.18296607  1.2829661  1.3829661

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.022594322 -0.022594322  0.07740568  0.27740568  0.87740568
#> 2  02 -2.689942269 -2.689942269 -2.58994227 -2.38994227 -1.78994227
#> 3  03  0.034182842  0.034182842  0.13418284  0.33418284  0.93418284
#> 4  04 -0.956440836 -0.956440836 -0.85644084 -0.65644084 -0.05644084
#> 5  05  0.177313880  0.177313880  0.27731388  0.47731388  1.07731388
#> 6  06  1.282624686  1.282624686  1.38262469  1.58262469  2.18262469
#> 7  07  0.839166484  0.839166484  0.93916648  1.13916648  1.73916648
#> 8  08  0.003576076  0.003576076  0.10357608  0.30357608  0.90357608
#> 9  09  0.273796695  0.273796695  0.37379669  0.57379669  1.17379669
#> 10 10 -0.311056006 -0.311056006 -0.21105601 -0.01105601  0.58894399