Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.98274797  0.98274797  1.08274797
#> 2  02  1.41529253  1.41529253  1.51529253
#> 3  03  1.40673424  1.40673424  1.50673424
#> 4  04  1.51246819  1.51246819  1.61246819
#> 5  05  0.24743750  0.24743750  0.34743750
#> 6  06 -0.12936130 -0.12936130 -0.02936130
#> 7  07  0.16068992  0.16068992  0.26068992
#> 8  08 -0.08722536 -0.08722536  0.01277464
#> 9  09 -0.34193769 -0.34193769 -0.24193769
#> 10 10  1.74252928  1.74252928  1.84252928

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01  1.34901566  1.34901566  1.449015659
#> 2  02 -0.10488776 -0.10488776 -0.004887765
#> 3  03 -0.38083809 -0.38083809 -0.280838089
#> 4  04 -1.75096370 -1.75096370 -1.650963702
#> 5  05  0.30086499  0.30086499  0.400864995
#> 6  06 -0.54201684 -0.54201684 -0.442016843
#> 7  07 -0.12915125 -0.12915125 -0.029151248
#> 8  08  0.02085299  0.02085299  0.120852993
#> 9  09  0.45196986  0.45196986  0.551969860
#> 10 10 -0.06908656 -0.06908656  0.030913435

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.33887724  1.4388772  1.53887724  1.63887724
#> 2  02  0.77834978  0.8783498  0.97834978  1.07834978
#> 3  03 -1.82133165 -1.7213316 -1.62133165 -1.52133165
#> 4  04 -0.62456757 -0.5245676 -0.42456757 -0.32456757
#> 5  05 -0.56215842 -0.4621584 -0.36215842 -0.26215842
#> 6  06  0.49598394  0.5959839  0.69598394  0.79598394
#> 7  07  0.92583900  1.0258390  1.12583900  1.22583900
#> 8  08 -0.23518933 -0.1351893 -0.03518933  0.06481067
#> 9  09  0.04220557  0.1422056  0.24220557  0.34220557
#> 10 10 -0.54914066 -0.4491407 -0.34914066 -0.24914066

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.0305206   1.0305206   1.1305206   1.3305206   1.9305206
#> 2  02  1.5722598   1.5722598   1.6722598   1.8722598   2.4722598
#> 3  03  0.5027011   0.5027011   0.6027011   0.8027011   1.4027011
#> 4  04  0.1153025   0.1153025   0.2153025   0.4153025   1.0153025
#> 5  05  0.6731035   0.6731035   0.7731035   0.9731035   1.5731035
#> 6  06  0.4760103   0.4760103   0.5760103   0.7760103   1.3760103
#> 7  07  0.5220124   0.5220124   0.6220124   0.8220124   1.4220124
#> 8  08  1.8512584   1.8512584   1.9512584   2.1512584   2.7512584
#> 9  09 -0.6978293  -0.6978293  -0.5978293  -0.3978293   0.2021707
#> 10 10  0.7643408   0.7643408   0.8643408   1.0643408   1.6643408