Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.28265157  1.28265157  1.3826516
#> 2  02  1.03314385  1.03314385  1.1331439
#> 3  03 -1.69430665 -1.69430665 -1.5943066
#> 4  04  0.03541141  0.03541141  0.1354114
#> 5  05 -0.41702723 -0.41702723 -0.3170272
#> 6  06 -2.16109625 -2.16109625 -2.0610963
#> 7  07 -0.80086055 -0.80086055 -0.7008605
#> 8  08  0.34759037  0.34759037  0.4475904
#> 9  09 -1.06669999 -1.06669999 -0.9667000
#> 10 10  0.51092121  0.51092121  0.6109212

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.4658231 -0.4658231 -0.3658231
#> 2  02  1.0208399  1.0208399  1.1208399
#> 3  03  1.6042883  1.6042883  1.7042883
#> 4  04 -2.5047961 -2.5047961 -2.4047961
#> 5  05 -0.4964756 -0.4964756 -0.3964756
#> 6  06 -1.7387924 -1.7387924 -1.6387924
#> 7  07  0.6838981  0.6838981  0.7838981
#> 8  08 -0.7658201 -0.7658201 -0.6658201
#> 9  09  0.7553068  0.7553068  0.8553068
#> 10 10  1.4944713  1.4944713  1.5944713

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.18852692  0.2885269  0.3885269  0.4885269
#> 2  02  0.08870660  0.1887066  0.2887066  0.3887066
#> 3  03  0.08718117  0.1871812  0.2871812  0.3871812
#> 4  04 -2.56848321 -2.4684832 -2.3684832 -2.2684832
#> 5  05  1.89542995  1.9954300  2.0954300  2.1954300
#> 6  06 -2.57802881 -2.4780288 -2.3780288 -2.2780288
#> 7  07  0.39622410  0.4962241  0.5962241  0.6962241
#> 8  08  0.17985631  0.2798563  0.3798563  0.4798563
#> 9  09  0.13976386  0.2397639  0.3397639  0.4397639
#> 10 10  0.59847833  0.6984783  0.7984783  0.8984783

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5963701   0.5963701  0.69637014  0.89637014   1.4963701
#> 2  02 -0.2602617  -0.2602617 -0.16026169  0.03973831   0.6397383
#> 3  03  0.2799463   0.2799463  0.37994631  0.57994631   1.1799463
#> 4  04  1.5530528   1.5530528  1.65305284  1.85305284   2.4530528
#> 5  05  0.2170867   0.2170867  0.31708669  0.51708669   1.1170867
#> 6  06 -0.4476263  -0.4476263 -0.34762634 -0.14762634   0.4523737
#> 7  07 -0.1945620  -0.1945620 -0.09456195  0.10543805   0.7054380
#> 8  08 -0.6365422  -0.6365422 -0.53654216 -0.33654216   0.2634578
#> 9  09 -1.8331823  -1.8331823 -1.73318228 -1.53318228  -0.9331823
#> 10 10  0.6687453   0.6687453  0.76874530  0.96874530   1.5687453