Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.7071217 -0.7071217 -0.6071217
#> 2  02  0.6375169  0.6375169  0.7375169
#> 3  03  1.1863232  1.1863232  1.2863232
#> 4  04  0.6705644  0.6705644  0.7705644
#> 5  05 -0.4386046 -0.4386046 -0.3386046
#> 6  06  0.4303831  0.4303831  0.5303831
#> 7  07  0.3430294  0.3430294  0.4430294
#> 8  08  0.2181479  0.2181479  0.3181479
#> 9  09 -0.7444374 -0.7444374 -0.6444374
#> 10 10  0.2493943  0.2493943  0.3493943

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.07717635 -1.07717635 -0.97717635
#> 2  02 -0.08187861 -0.08187861  0.01812139
#> 3  03 -1.41392878 -1.41392878 -1.31392878
#> 4  04  0.85202736  0.85202736  0.95202736
#> 5  05 -0.60759583 -0.60759583 -0.50759583
#> 6  06 -0.52069297 -0.52069297 -0.42069297
#> 7  07  0.81239622  0.81239622  0.91239622
#> 8  08 -1.37807176 -1.37807176 -1.27807176
#> 9  09 -0.85828896 -0.85828896 -0.75828896
#> 10 10  0.71397101  0.71397101  0.81397101

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.7123912  0.81239116  0.91239116  1.01239116
#> 2  02 -0.4778290 -0.37782904 -0.27782904 -0.17782904
#> 3  03 -0.9595296 -0.85952963 -0.75952963 -0.65952963
#> 4  04 -0.3743356 -0.27433556 -0.17433556 -0.07433556
#> 5  05 -1.1223815 -1.02238153 -0.92238153 -0.82238153
#> 6  06  0.4394823  0.53948229  0.63948229  0.73948229
#> 7  07 -0.2149160 -0.11491600 -0.01491600  0.08508400
#> 8  08 -0.1170742 -0.01707415  0.08292585  0.18292585
#> 9  09  0.1493295  0.24932952  0.34932952  0.44932952
#> 10 10  0.1296121  0.22961208  0.32961208  0.42961208

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.0276379   0.0276379   0.1276379   0.3276379   0.9276379
#> 2  02 -0.4221510  -0.4221510  -0.3221510  -0.1221510   0.4778490
#> 3  03  1.5186041   1.5186041   1.6186041   1.8186041   2.4186041
#> 4  04  0.5747315   0.5747315   0.6747315   0.8747315   1.4747315
#> 5  05  0.5939841   0.5939841   0.6939841   0.8939841   1.4939841
#> 6  06  1.6934730   1.6934730   1.7934730   1.9934730   2.5934730
#> 7  07 -1.9788271  -1.9788271  -1.8788271  -1.6788271  -1.0788271
#> 8  08 -2.2321211  -2.2321211  -2.1321211  -1.9321211  -1.3321211
#> 9  09 -0.6819304  -0.6819304  -0.5819304  -0.3819304   0.2180696
#> 10 10  0.2664603   0.2664603   0.3664603   0.5664603   1.1664603