Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.2887739  1.2887739  1.3887739
#> 2  02 -1.3148389 -1.3148389 -1.2148389
#> 3  03  0.4963348  0.4963348  0.5963348
#> 4  04  1.1307654  1.1307654  1.2307654
#> 5  05 -0.8923251 -0.8923251 -0.7923251
#> 6  06 -0.4537870 -0.4537870 -0.3537870
#> 7  07  1.7371847  1.7371847  1.8371847
#> 8  08  1.1505129  1.1505129  1.2505129
#> 9  09  0.6176754  0.6176754  0.7176754
#> 10 10  1.2771062  1.2771062  1.3771062

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.26444152  1.26444152  1.36444152
#> 2  02  0.59481098  0.59481098  0.69481098
#> 3  03  0.52254256  0.52254256  0.62254256
#> 4  04  0.03103255  0.03103255  0.13103255
#> 5  05 -0.16139899 -0.16139899 -0.06139899
#> 6  06  0.28699772  0.28699772  0.38699772
#> 7  07  0.32747663  0.32747663  0.42747663
#> 8  08  0.83437409  0.83437409  0.93437409
#> 9  09 -0.15148690 -0.15148690 -0.05148690
#> 10 10 -0.71615093 -0.71615093 -0.61615093

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.38151117 -0.28151117 -0.18151117 -0.08151117
#> 2  02 -1.12665264 -1.02665264 -0.92665264 -0.82665264
#> 3  03  0.76299953  0.86299953  0.96299953  1.06299953
#> 4  04 -0.31738816 -0.21738816 -0.11738816 -0.01738816
#> 5  05 -1.23706606 -1.13706606 -1.03706606 -0.93706606
#> 6  06  0.29252648  0.39252648  0.49252648  0.59252648
#> 7  07 -0.31174609 -0.21174609 -0.11174609 -0.01174609
#> 8  08 -0.05165943  0.04834057  0.14834057  0.24834057
#> 9  09 -0.26993978 -0.16993978 -0.06993978  0.03006022
#> 10 10 -2.24971828 -2.14971828 -2.04971828 -1.94971828

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -3.09772232 -3.09772232  -2.9977223 -2.79772232  -2.1977223
#> 2  02  1.66116367  1.66116367   1.7611637  1.96116367   2.5611637
#> 3  03 -0.23095205 -0.23095205  -0.1309520  0.06904795   0.6690480
#> 4  04 -0.69895569 -0.69895569  -0.5989557 -0.39895569   0.2010443
#> 5  05 -0.54885116 -0.54885116  -0.4488512 -0.24885116   0.3511488
#> 6  06  0.67198615  0.67198615   0.7719861  0.97198615   1.5719861
#> 7  07  0.04535897  0.04535897   0.1453590  0.34535897   0.9453590
#> 8  08  1.11274413  1.11274413   1.2127441  1.41274413   2.0127441
#> 9  09 -0.78431050 -0.78431050  -0.6843105 -0.48431050   0.1156895
#> 10 10  1.25259385  1.25259385   1.3525939  1.55259385   2.1525939