Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.2018856 -1.2018856 -1.1018856
#> 2  02  0.9231016  0.9231016  1.0231016
#> 3  03  1.7925831  1.7925831  1.8925831
#> 4  04 -0.7140753 -0.7140753 -0.6140753
#> 5  05  0.3624475  0.3624475  0.4624475
#> 6  06 -1.7520685 -1.7520685 -1.6520685
#> 7  07  0.2854402  0.2854402  0.3854402
#> 8  08 -1.2338332 -1.2338332 -1.1338332
#> 9  09  0.2764585  0.2764585  0.3764585
#> 10 10  0.3722117  0.3722117  0.4722117

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.5522177 -0.5522177 -0.45221773
#> 2  02  1.1186282  1.1186282  1.21862823
#> 3  03 -1.5900091 -1.5900091 -1.49000914
#> 4  04 -0.1192068 -0.1192068 -0.01920678
#> 5  05 -1.7302464 -1.7302464 -1.63024636
#> 6  06 -0.9707878 -0.9707878 -0.87078781
#> 7  07  0.5780227  0.5780227  0.67802268
#> 8  08 -0.2752844 -0.2752844 -0.17528440
#> 9  09  1.0452420  1.0452420  1.14524203
#> 10 10 -0.1673868 -0.1673868 -0.06738678

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.06378110 -0.9637811 -0.8637811 -0.7637811
#> 2  02  1.64799915  1.7479992  1.8479992  1.9479992
#> 3  03  0.46310147  0.5631015  0.6631015  0.7631015
#> 4  04 -0.79729301 -0.6972930 -0.5972930 -0.4972930
#> 5  05 -0.37802270 -0.2780227 -0.1780227 -0.0780227
#> 6  06  0.02112379  0.1211238  0.2211238  0.3211238
#> 7  07  0.33918722  0.4391872  0.5391872  0.6391872
#> 8  08  1.14737530  1.2473753  1.3473753  1.4473753
#> 9  09 -1.97623959 -1.8762396 -1.7762396 -1.6762396
#> 10 10  0.86615429  0.9661543  1.0661543  1.1661543

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.95226977 -0.95226977 -0.85226977  -0.6522698 -0.05226977
#> 2  02  1.07835915  1.07835915  1.17835915   1.3783592  1.97835915
#> 3  03  0.86396619  0.86396619  0.96396619   1.1639662  1.76396619
#> 4  04 -0.07791835 -0.07791835  0.02208165   0.2220817  0.82208165
#> 5  05  1.15361660  1.15361660  1.25361660   1.4536166  2.05361660
#> 6  06  0.02312865  0.02312865  0.12312865   0.3231287  0.92312865
#> 7  07 -0.59039226 -0.59039226 -0.49039226  -0.2903923  0.30960774
#> 8  08 -1.00429173 -1.00429173 -0.90429173  -0.7042917 -0.10429173
#> 9  09  0.50424617  0.50424617  0.60424617   0.8042462  1.40424617
#> 10 10  0.32631742  0.32631742  0.42631742   0.6263174  1.22631742