Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.947971075  0.947971075  1.04797107
#> 2  02  1.371968922  1.371968922  1.47196892
#> 3  03 -0.006267188 -0.006267188  0.09373281
#> 4  04 -0.186252907 -0.186252907 -0.08625291
#> 5  05  0.147592994  0.147592994  0.24759299
#> 6  06  1.611730031  1.611730031  1.71173003
#> 7  07  0.077429664  0.077429664  0.17742966
#> 8  08  0.858513020  0.858513020  0.95851302
#> 9  09 -0.689897798 -0.689897798 -0.58989780
#> 10 10  0.158174812  0.158174812  0.25817481

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.8908599 -1.8908599 -1.7908599
#> 2  02  0.2528610  0.2528610  0.3528610
#> 3  03  0.3915291  0.3915291  0.4915291
#> 4  04 -1.7239107 -1.7239107 -1.6239107
#> 5  05 -1.6754822 -1.6754822 -1.5754822
#> 6  06  2.2600459  2.2600459  2.3600459
#> 7  07  1.2180883  1.2180883  1.3180883
#> 8  08 -1.1468726 -1.1468726 -1.0468726
#> 9  09 -0.7761918 -0.7761918 -0.6761918
#> 10 10  0.3052001  0.3052001  0.4052001

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.446913035  0.54691303  0.64691303  0.74691303
#> 2  02  1.163458468  1.26345847  1.36345847  1.46345847
#> 3  03 -0.996348605 -0.89634860 -0.79634860 -0.69634860
#> 4  04 -0.873405180 -0.77340518 -0.67340518 -0.57340518
#> 5  05  0.004104354  0.10410435  0.20410435  0.30410435
#> 6  06 -0.828395952 -0.72839595 -0.62839595 -0.52839595
#> 7  07 -0.844933000 -0.74493300 -0.64493300 -0.54493300
#> 8  08 -0.325124921 -0.22512492 -0.12512492 -0.02512492
#> 9  09  0.818102839  0.91810284  1.01810284  1.11810284
#> 10 10 -0.162187660 -0.06218766  0.03781234  0.13781234

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.1898273  -1.1898273 -1.08982726  -0.8898273  -0.2898273
#> 2  02 -0.5503094  -0.5503094 -0.45030938  -0.2503094   0.3496906
#> 3  03  2.8743671   2.8743671  2.97436705   3.1743671   3.7743671
#> 4  04 -0.7062455  -0.7062455 -0.60624552  -0.4062455   0.1937545
#> 5  05 -1.0203797  -1.0203797 -0.92037971  -0.7203797  -0.1203797
#> 6  06 -0.7981427  -0.7981427 -0.69814269  -0.4981427   0.1018573
#> 7  07  0.1957041   0.1957041  0.29570406   0.4957041   1.0957041
#> 8  08 -0.1977425  -0.1977425 -0.09774248   0.1022575   0.7022575
#> 9  09 -1.9009172  -1.9009172 -1.80091722  -1.6009172  -1.0009172
#> 10 10  1.5589063   1.5589063  1.65890626   1.8589063   2.4589063