Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.02960752 -0.02960752  0.07039248
#> 2  02 -1.52293394 -1.52293394 -1.42293394
#> 3  03 -1.04154453 -1.04154453 -0.94154453
#> 4  04 -0.24599578 -0.24599578 -0.14599578
#> 5  05  0.07161092  0.07161092  0.17161092
#> 6  06  0.82001205  0.82001205  0.92001205
#> 7  07  0.09202364  0.09202364  0.19202364
#> 8  08 -0.74577886 -0.74577886 -0.64577886
#> 9  09 -0.54116680 -0.54116680 -0.44116680
#> 10 10  1.55844301  1.55844301  1.65844301

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.420360132  0.420360132  0.52036013
#> 2  02 -0.778022475 -0.778022475 -0.67802248
#> 3  03 -0.005796977 -0.005796977  0.09420302
#> 4  04  0.181329762  0.181329762  0.28132976
#> 5  05 -1.262889697 -1.262889697 -1.16288970
#> 6  06  1.748005470  1.748005470  1.84800547
#> 7  07 -0.369177525 -0.369177525 -0.26917752
#> 8  08  0.134761921  0.134761921  0.23476192
#> 9  09  0.285525669  0.285525669  0.38552567
#> 10 10  0.488480953  0.488480953  0.58848095

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  1.19892663  1.29892663  1.39892663  1.4989266
#> 2  02  0.51546672  0.61546672  0.71546672  0.8154667
#> 3  03 -1.50736733 -1.40736733 -1.30736733 -1.2073673
#> 4  04 -0.14949845 -0.04949845  0.05050155  0.1505015
#> 5  05 -0.72700479 -0.62700479 -0.52700479 -0.4270048
#> 6  06 -0.80885965 -0.70885965 -0.60885965 -0.5088596
#> 7  07 -1.21279211 -1.11279211 -1.01279211 -0.9127921
#> 8  08  0.38021272  0.48021272  0.58021272  0.6802127
#> 9  09  1.13751975  1.23751975  1.33751975  1.4375198
#> 10 10 -0.02710952  0.07289048  0.17289048  0.2728905

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.3344421  -0.3344421  -0.2344421 -0.03444212  0.56555788
#> 2  02  0.7729629   0.7729629   0.8729629  1.07296290  1.67296290
#> 3  03 -0.6215234  -0.6215234  -0.5215234 -0.32152342  0.27847658
#> 4  04  1.0541488   1.0541488   1.1541488  1.35414884  1.95414884
#> 5  05 -0.7770986  -0.7770986  -0.6770986 -0.47709861  0.12290139
#> 6  06  0.6476943   0.6476943   0.7476943  0.94769428  1.54769428
#> 7  07 -0.8033232  -0.8033232  -0.7033232 -0.50332317  0.09667683
#> 8  08 -0.5997866  -0.5997866  -0.4997866 -0.29978656  0.30021344
#> 9  09  0.7854809   0.7854809   0.8854809  1.08548095  1.68548095
#> 10 10  2.2584263   2.2584263   2.3584263  2.55842626  3.15842626