Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.9694358 -0.9694358 -0.86943584
#> 2  02 -1.4151503 -1.4151503 -1.31515026
#> 3  03  0.5201601  0.5201601  0.62016010
#> 4  04  0.6815373  0.6815373  0.78153733
#> 5  05 -0.1567041 -0.1567041 -0.05670406
#> 6  06 -0.2300251 -0.2300251 -0.13002515
#> 7  07  1.7218028  1.7218028  1.82180275
#> 8  08  1.7375192  1.7375192  1.83751920
#> 9  09 -0.1741072 -0.1741072 -0.07410719
#> 10 10  1.4217120  1.4217120  1.52171196

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  0.602821640  0.602821640  0.70282164
#> 2  02  0.763669819  0.763669819  0.86366982
#> 3  03  0.577518290  0.577518290  0.67751829
#> 4  04 -0.519639871 -0.519639871 -0.41963987
#> 5  05 -0.330413418 -0.330413418 -0.23041342
#> 6  06 -0.005134814 -0.005134814  0.09486519
#> 7  07 -0.423610448 -0.423610448 -0.32361045
#> 8  08 -0.550217040 -0.550217040 -0.45021704
#> 9  09  1.459545042  1.459545042  1.55954504
#> 10 10 -2.573992619 -2.573992619 -2.47399262

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  1.281975052  1.38197505  1.4819751  1.58197505
#> 2  02  0.007950943  0.10795094  0.2079509  0.30795094
#> 3  03 -0.016148774  0.08385123  0.1838512  0.28385123
#> 4  04 -1.604645970 -1.50464597 -1.4046460 -1.30464597
#> 5  05 -0.377080544 -0.27708054 -0.1770805 -0.07708054
#> 6  06 -1.544007614 -1.44400761 -1.3440076 -1.24400761
#> 7  07  0.211439013  0.31143901  0.4114390  0.51143901
#> 8  08  0.783770074  0.88377007  0.9837701  1.08377007
#> 9  09 -1.842289341 -1.74228934 -1.6422893 -1.54228934
#> 10 10 -0.910091063 -0.81009106 -0.7100911 -0.61009106

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  2.0589688   2.0589688   2.1589688  2.35896884   2.9589688
#> 2  02  0.7955072   0.7955072   0.8955072  1.09550723   1.6955072
#> 3  03  0.3759393   0.3759393   0.4759393  0.67593931   1.2759393
#> 4  04  0.7041702   0.7041702   0.8041702  1.00417024   1.6041702
#> 5  05  2.2023987   2.2023987   2.3023987  2.50239870   3.1023987
#> 6  06  0.9476278   0.9476278   1.0476278  1.24762779   1.8476278
#> 7  07  0.1499003   0.1499003   0.2499003  0.44990027   1.0499003
#> 8  08 -0.3571528  -0.3571528  -0.2571528 -0.05715282   0.5428472
#> 9  09  2.3350893   2.3350893   2.4350893  2.63508930   3.2350893
#> 10 10  0.1076672   0.1076672   0.2076672  0.40766724   1.0076672