Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.0544380 -0.0544380  0.0455620
#> 2  02 -0.8658720 -0.8658720 -0.7658720
#> 3  03 -0.2436019 -0.2436019 -0.1436019
#> 4  04  0.2216456  0.2216456  0.3216456
#> 5  05  1.3401390  1.3401390  1.4401390
#> 6  06  0.4777975  0.4777975  0.5777975
#> 7  07  0.8093038  0.8093038  0.9093038
#> 8  08 -0.4772554 -0.4772554 -0.3772554
#> 9  09  0.8788002  0.8788002  0.9788002
#> 10 10  0.2829996  0.2829996  0.3829996

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -2.0787948 -2.0787948 -1.9787948
#> 2  02  2.1327637  2.1327637  2.2327637
#> 3  03  0.3557581  0.3557581  0.4557581
#> 4  04  0.5393171  0.5393171  0.6393171
#> 5  05  0.0716819  0.0716819  0.1716819
#> 6  06 -0.6046148 -0.6046148 -0.5046148
#> 7  07 -0.9755641 -0.9755641 -0.8755641
#> 8  08  1.0129662  1.0129662  1.1129662
#> 9  09  0.3386750  0.3386750  0.4386750
#> 10 10 -0.3771610 -0.3771610 -0.2771610

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.3671444  0.4671444  0.56714443  0.66714443
#> 2  02  0.2434823  0.3434823  0.44348230  0.54348230
#> 3  03  2.4160259  2.5160259  2.61602592  2.71602592
#> 4  04 -1.4164141 -1.3164141 -1.21641412 -1.11641412
#> 5  05  0.8152115  0.9152115  1.01521149  1.11521149
#> 6  06  1.2224837  1.3224837  1.42248366  1.52248366
#> 7  07 -0.2418967 -0.1418967 -0.04189669  0.05810331
#> 8  08  1.2025063  1.3025063  1.40250627  1.50250627
#> 9  09 -1.6238914 -1.5238914 -1.42389136 -1.32389136
#> 10 10 -0.2706973 -0.1706973 -0.07069732  0.02930268

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.6887056  -0.6887056  -0.5887056 -0.38870559  0.21129441
#> 2  02 -0.9870620  -0.9870620  -0.8870620 -0.68706204 -0.08706204
#> 3  03 -0.2767879  -0.2767879  -0.1767879  0.02321206  0.62321206
#> 4  04 -1.2207258  -1.2207258  -1.1207258 -0.92072577 -0.32072577
#> 5  05 -0.3102834  -0.3102834  -0.2102834 -0.01028343  0.58971657
#> 6  06 -1.3659951  -1.3659951  -1.2659951 -1.06599506 -0.46599506
#> 7  07  2.1819762   2.1819762   2.2819762  2.48197619  3.08197619
#> 8  08  1.3463020   1.3463020   1.4463020  1.64630196  2.24630196
#> 9  09  0.8789427   0.8789427   0.9789427  1.17894272  1.77894272
#> 10 10  0.2245760   0.2245760   0.3245760  0.52457598  1.12457598