Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID             U         Y_Z_0       Y_Z_1
#> 1  01 -0.1742016751 -0.1742016751 -0.07420168
#> 2  02 -0.4562213796 -0.4562213796 -0.35622138
#> 3  03 -0.9890653826 -0.9890653826 -0.88906538
#> 4  04 -1.2830216036 -1.2830216036 -1.18302160
#> 5  05  1.2300179653  1.2300179653  1.33001797
#> 6  06  0.0679318281  0.0679318281  0.16793183
#> 7  07  0.3962848767  0.3962848767  0.49628488
#> 8  08  0.9908567194  0.9908567194  1.09085672
#> 9  09 -0.0004412112 -0.0004412112  0.09955879
#> 10 10 -0.3999767861 -0.3999767861 -0.29997679

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.7003769 -1.7003769 -1.6003769
#> 2  02 -0.5198832 -0.5198832 -0.4198832
#> 3  03 -1.3639592 -1.3639592 -1.2639592
#> 4  04 -0.6966926 -0.6966926 -0.5966926
#> 5  05  0.5700802  0.5700802  0.6700802
#> 6  06  0.3833108  0.3833108  0.4833108
#> 7  07  1.7983751  1.7983751  1.8983751
#> 8  08  2.2181728  2.2181728  2.3181728
#> 9  09  1.3832527  1.3832527  1.4832527
#> 10 10  0.2622051  0.2622051  0.3622051

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.80286174  0.90286174  1.00286174  1.10286174
#> 2  02 -0.69537177 -0.59537177 -0.49537177 -0.39537177
#> 3  03  1.55745823  1.65745823  1.75745823  1.85745823
#> 4  04  0.06624074  0.16624074  0.26624074  0.36624074
#> 5  05 -1.39354010 -1.29354010 -1.19354010 -1.09354010
#> 6  06  1.14429659  1.24429659  1.34429659  1.44429659
#> 7  07  1.14269085  1.24269085  1.34269085  1.44269085
#> 8  08 -0.05844477  0.04155523  0.14155523  0.24155523
#> 9  09 -0.22527303 -0.12527303 -0.02527303  0.07472697
#> 10 10 -1.15123512 -1.05123512 -0.95123512 -0.85123512

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5892129   0.5892129   0.6892129  0.88921294  1.48921294
#> 2  02 -1.2688319  -1.2688319  -1.1688319 -0.96883189 -0.36883189
#> 3  03  0.7942281   0.7942281   0.8942281  1.09422806  1.69422806
#> 4  04 -0.6915354  -0.6915354  -0.5915354 -0.39153537  0.20846463
#> 5  05  1.1230055   1.1230055   1.2230055  1.42300551  2.02300551
#> 6  06  0.6935144   0.6935144   0.7935144  0.99351435  1.59351435
#> 7  07 -1.0267941  -1.0267941  -0.9267941 -0.72679409 -0.12679409
#> 8  08  1.5463192   1.5463192   1.6463192  1.84631923  2.44631923
#> 9  09 -0.8041756  -0.8041756  -0.7041756 -0.50417564  0.09582436
#> 10 10 -0.3856716  -0.3856716  -0.2856716 -0.08567157  0.51432843