Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1155433 -0.1155433 -0.01554329
#> 2  02 -1.0173241 -1.0173241 -0.91732408
#> 3  03  0.4484447  0.4484447  0.54844467
#> 4  04 -1.0359014 -1.0359014 -0.93590143
#> 5  05  0.5282553  0.5282553  0.62825526
#> 6  06 -0.9165271 -0.9165271 -0.81652712
#> 7  07 -0.1945937 -0.1945937 -0.09459369
#> 8  08 -0.2396296 -0.2396296 -0.13962956
#> 9  09  0.2820335  0.2820335  0.38203351
#> 10 10 -2.1410541 -2.1410541 -2.04105407

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -1.33111276 -1.33111276 -1.2311128
#> 2  02  0.27892469  0.27892469  0.3789247
#> 3  03  0.03028114  0.03028114  0.1302811
#> 4  04 -0.44057819 -0.44057819 -0.3405782
#> 5  05  0.29369618  0.29369618  0.3936962
#> 6  06 -0.21121357 -0.21121357 -0.1112136
#> 7  07  0.22144524  0.22144524  0.3214452
#> 8  08  0.37665073  0.37665073  0.4766507
#> 9  09 -0.84127247 -0.84127247 -0.7412725
#> 10 10 -0.94229214 -0.94229214 -0.8422921

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.2788056  0.37880563  0.47880563  0.57880563
#> 2  02 -0.5841597 -0.48415971 -0.38415971 -0.28415971
#> 3  03  1.2247356  1.32473555  1.42473555  1.52473555
#> 4  04 -0.1579408 -0.05794078  0.04205922  0.14205922
#> 5  05 -1.1175630 -1.01756299 -0.91756299 -0.81756299
#> 6  06 -0.7347523 -0.63475229 -0.53475229 -0.43475229
#> 7  07 -0.1697113 -0.06971133  0.03028867  0.13028867
#> 8  08 -0.4303557 -0.33035569 -0.23035569 -0.13035569
#> 9  09 -1.2255060 -1.12550597 -1.02550597 -0.92550597
#> 10 10 -0.2544680 -0.15446804 -0.05446804  0.04553196

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1   Y_Z1_1_Z2_1
#> 1  01 -0.2148324  -0.2148324  -0.1148324  0.08516756  0.6851675570
#> 2  02 -0.8997221  -0.8997221  -0.7997221 -0.59972213  0.0002778711
#> 3  03  0.6414268   0.6414268   0.7414268  0.94142682  1.5414268236
#> 4  04  0.4827154   0.4827154   0.5827154  0.78271536  1.3827153633
#> 5  05 -0.4403928  -0.4403928  -0.3403928 -0.14039280  0.4596071975
#> 6  06 -1.0509066  -1.0509066  -0.9509066 -0.75090664 -0.1509066418
#> 7  07  0.6443000   0.6443000   0.7443000  0.94430002  1.5443000188
#> 8  08 -0.2640145  -0.2640145  -0.1640145  0.03598546  0.6359854582
#> 9  09  1.1647812   1.1647812   1.2647812  1.46478122  2.0647812219
#> 10 10  0.2320162   0.2320162   0.3320162  0.53201622  1.1320162220