Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.4857303  0.4857303  0.5857303
#> 2  02 -1.0618779 -1.0618779 -0.9618779
#> 3  03  0.6069372  0.6069372  0.7069372
#> 4  04 -0.3292366 -0.3292366 -0.2292366
#> 5  05  0.4758546  0.4758546  0.5758546
#> 6  06 -2.2133230 -2.2133230 -2.1133230
#> 7  07 -1.0018445 -1.0018445 -0.9018445
#> 8  08  0.3698502  0.3698502  0.4698502
#> 9  09 -0.5746685 -0.5746685 -0.4746685
#> 10 10  0.5951532  0.5951532  0.6951532

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.0864531  1.0864531  1.1864531
#> 2  02  0.1849706  0.1849706  0.2849706
#> 3  03 -0.9530594 -0.9530594 -0.8530594
#> 4  04  0.2851873  0.2851873  0.3851873
#> 5  05 -2.4183438 -2.4183438 -2.3183438
#> 6  06 -1.0185815 -1.0185815 -0.9185815
#> 7  07  1.0397577  1.0397577  1.1397577
#> 8  08  1.2248360  1.2248360  1.3248360
#> 9  09 -0.7621083 -0.7621083 -0.6621083
#> 10 10  1.6455735  1.6455735  1.7455735

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -1.22695172 -1.12695172 -1.02695172 -0.9269517
#> 2  02 -1.53916738 -1.43916738 -1.33916738 -1.2391674
#> 3  03  2.07134097  2.17134097  2.27134097  2.3713410
#> 4  04 -0.03535333  0.06464667  0.16464667  0.2646467
#> 5  05 -1.53480579 -1.43480579 -1.33480579 -1.2348058
#> 6  06  0.65137881  0.75137881  0.85137881  0.9513788
#> 7  07  1.51744773  1.61744773  1.71744773  1.8174477
#> 8  08 -0.14462995 -0.04462995  0.05537005  0.1553701
#> 9  09  0.05365873  0.15365873  0.25365873  0.3536587
#> 10 10  0.74386868  0.84386868  0.94386868  1.0438687

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.6501093  -0.6501093  -0.5501093 -0.35010927   0.2498907
#> 2  02  0.4310502   0.4310502   0.5310502  0.73105017   1.3310502
#> 3  03 -0.6199620  -0.6199620  -0.5199620 -0.31996203   0.2800380
#> 4  04  0.1923053   0.1923053   0.2923053  0.49230527   1.0923053
#> 5  05  1.4652986   1.4652986   1.5652986  1.76529859   2.3652986
#> 6  06 -0.6523667  -0.6523667  -0.5523667 -0.35236667   0.2476333
#> 7  07 -2.0715282  -2.0715282  -1.9715282 -1.77152820  -1.1715282
#> 8  08 -0.3547687  -0.3547687  -0.2547687 -0.05476873   0.5452313
#> 9  09 -1.2483415  -1.2483415  -1.1483415 -0.94834145  -0.3483415
#> 10 10  0.2283962   0.2283962   0.3283962  0.52839623   1.1283962