Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.6790357 -0.6790357 -0.5790357
#> 2  02 -1.0553217 -1.0553217 -0.9553217
#> 3  03  1.3459668  1.3459668  1.4459668
#> 4  04  1.7622815  1.7622815  1.8622815
#> 5  05  1.7941460  1.7941460  1.8941460
#> 6  06  0.1628493  0.1628493  0.2628493
#> 7  07  0.3025824  0.3025824  0.4025824
#> 8  08  0.5357770  0.5357770  0.6357770
#> 9  09 -0.5286681 -0.5286681 -0.4286681
#> 10 10  1.3516427  1.3516427  1.4516427

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01  0.524036517  0.524036517  0.6240365
#> 2  02  0.941702441  0.941702441  1.0417024
#> 3  03  0.899628293  0.899628293  0.9996283
#> 4  04  1.310811214  1.310811214  1.4108112
#> 5  05  1.133677618  1.133677618  1.2336776
#> 6  06 -0.005470998 -0.005470998  0.0945290
#> 7  07 -0.390902153 -0.390902153 -0.2909022
#> 8  08  0.458235058  0.458235058  0.5582351
#> 9  09  0.545186449  0.545186449  0.6451864
#> 10 10 -0.253049495 -0.253049495 -0.1530495

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1        Y_Z_2       Y_Z_3
#> 1  01 -1.0609384 -0.9609384 -0.860938383 -0.76093838
#> 2  02  1.5938187  1.6938187  1.793818748  1.89381875
#> 3  03 -0.2073956 -0.1073956 -0.007395634  0.09260437
#> 4  04  1.2664791  1.3664791  1.466479051  1.56647905
#> 5  05 -1.7027146 -1.6027146 -1.502714582 -1.40271458
#> 6  06 -1.2205500 -1.1205500 -1.020549976 -0.92054998
#> 7  07  0.9875920  1.0875920  1.187592014  1.28759201
#> 8  08 -0.4090768 -0.3090768 -0.209076763 -0.10907676
#> 9  09 -0.5749095 -0.4749095 -0.374909472 -0.27490947
#> 10 10  0.5869648  0.6869648  0.786964753  0.88696475

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID             U   Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.8141918071 -1.8141918071 -1.71419181 -1.51419181  -0.9141918
#> 2  02 -0.3792458175 -0.3792458175 -0.27924582 -0.07924582   0.5207542
#> 3  03 -0.5915105386 -0.5915105386 -0.49151054 -0.29151054   0.3084895
#> 4  04  0.8183902693  0.8183902693  0.91839027  1.11839027   1.7183903
#> 5  05 -0.5241241422 -0.5241241422 -0.42412414 -0.22412414   0.3758759
#> 6  06  0.2625845559  0.2625845559  0.36258456  0.56258456   1.1625846
#> 7  07  0.7969783435  0.7969783435  0.89697834  1.09697834   1.6969783
#> 8  08 -0.0002842826 -0.0002842826  0.09971572  0.29971572   0.8997157
#> 9  09  0.2244693488  0.2244693488  0.32446935  0.52446935   1.1244693
#> 10 10 -1.4029881619 -1.4029881619 -1.30298816 -1.10298816  -0.5029882