Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.43808468 -0.43808468 -0.33808468
#> 2  02 -0.08496033 -0.08496033  0.01503967
#> 3  03 -1.65219931 -1.65219931 -1.55219931
#> 4  04 -1.02151714 -1.02151714 -0.92151714
#> 5  05  1.08455989  1.08455989  1.18455989
#> 6  06  1.06359774  1.06359774  1.16359774
#> 7  07  1.56520928  1.56520928  1.66520928
#> 8  08 -1.18407747 -1.18407747 -1.08407747
#> 9  09  0.93943924  0.93943924  1.03943924
#> 10 10 -0.58611065 -0.58611065 -0.48611065

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.6691975  0.6691975  0.7691975
#> 2  02  0.2738540  0.2738540  0.3738540
#> 3  03  1.1100691  1.1100691  1.2100691
#> 4  04  2.0323981  2.0323981  2.1323981
#> 5  05  0.8919596  0.8919596  0.9919596
#> 6  06  0.6696164  0.6696164  0.7696164
#> 7  07 -1.1056134 -1.1056134 -1.0056134
#> 8  08 -2.9867661 -2.9867661 -2.8867661
#> 9  09 -3.2551876 -3.2551876 -3.1551876
#> 10 10 -1.5462770 -1.5462770 -1.4462770

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.8260840  0.9260840  1.0260840  1.1260840
#> 2  02 -0.7119781 -0.6119781 -0.5119781 -0.4119781
#> 3  03  0.2831395  0.3831395  0.4831395  0.5831395
#> 4  04 -0.6827135 -0.5827135 -0.4827135 -0.3827135
#> 5  05 -0.4212410 -0.3212410 -0.2212410 -0.1212410
#> 6  06  1.7008349  1.8008349  1.9008349  2.0008349
#> 7  07  1.5682238  1.6682238  1.7682238  1.8682238
#> 8  08  1.0959201  1.1959201  1.2959201  1.3959201
#> 9  09  0.1000405  0.2000405  0.3000405  0.4000405
#> 10 10  0.9204174  1.0204174  1.1204174  1.2204174

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -2.02204827 -2.02204827 -1.92204827 -1.72204827  -1.1220483
#> 2  02 -0.28866227 -0.28866227 -0.18866227  0.01133773   0.6113377
#> 3  03  0.89079590  0.89079590  0.99079590  1.19079590   1.7907959
#> 4  04 -0.03312409 -0.03312409  0.06687591  0.26687591   0.8668759
#> 5  05 -1.06827990 -1.06827990 -0.96827990 -0.76827990  -0.1682799
#> 6  06  0.83688148  0.83688148  0.93688148  1.13688148   1.7368815
#> 7  07 -0.76342623 -0.76342623 -0.66342623 -0.46342623   0.1365738
#> 8  08 -2.23698909 -2.23698909 -2.13698909 -1.93698909  -1.3369891
#> 9  09  1.35133169  1.35133169  1.45133169  1.65133169   2.2513317
#> 10 10  0.25435359  0.25435359  0.35435359  0.55435359   1.1543536