Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.7398885 -0.7398885 -0.6398885
#> 2  02  0.3920474  0.3920474  0.4920474
#> 3  03 -0.9385095 -0.9385095 -0.8385095
#> 4  04 -1.5032129 -1.5032129 -1.4032129
#> 5  05 -0.2788640 -0.2788640 -0.1788640
#> 6  06 -1.7416561 -1.7416561 -1.6416561
#> 7  07  0.7134054  0.7134054  0.8134054
#> 8  08 -0.7965443 -0.7965443 -0.6965443
#> 9  09 -0.2152345 -0.2152345 -0.1152345
#> 10 10 -0.9897154 -0.9897154 -0.8897154

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  2.6305335  2.6305335  2.7305335
#> 2  02  0.3835996  0.3835996  0.4835996
#> 3  03  1.3885050  1.3885050  1.4885050
#> 4  04  0.1086552  0.1086552  0.2086552
#> 5  05 -0.2048054 -0.2048054 -0.1048054
#> 6  06  0.6654746  0.6654746  0.7654746
#> 7  07  0.6834922  0.6834922  0.7834922
#> 8  08 -0.8759230 -0.8759230 -0.7759230
#> 9  09  0.3703129  0.3703129  0.4703129
#> 10 10 -1.6585707 -1.6585707 -1.5585707

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1        Y_Z_2       Y_Z_3
#> 1  01 -1.56963716 -1.4696372 -1.369637162 -1.26963716
#> 2  02 -0.33870171 -0.2387017 -0.138701710 -0.03870171
#> 3  03  1.37264463  1.4726446  1.572644628  1.67264463
#> 4  04 -0.45253969 -0.3525397 -0.252539688 -0.15253969
#> 5  05  0.55106323  0.6510632  0.751063233  0.85106323
#> 6  06 -0.77528169 -0.6752817 -0.575281691 -0.47528169
#> 7  07 -1.05046707 -0.9504671 -0.850467069 -0.75046707
#> 8  08 -0.20352421 -0.1035242 -0.003524215  0.09647579
#> 9  09  0.04979984  0.1497998  0.249799838  0.34979984
#> 10 10 -0.74886100 -0.6488610 -0.548861001 -0.44886100

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  2.09793060  2.09793060  2.19793060  2.39793060   2.9979306
#> 2  02  0.92920155  0.92920155  1.02920155  1.22920155   1.8292015
#> 3  03 -0.38458675 -0.38458675 -0.28458675 -0.08458675   0.5154132
#> 4  04 -0.01032501 -0.01032501  0.08967499  0.28967499   0.8896750
#> 5  05  0.49359867  0.49359867  0.59359867  0.79359867   1.3935987
#> 6  06  1.52525826  1.52525826  1.62525826  1.82525826   2.4252583
#> 7  07 -1.65511174 -1.65511174 -1.55511174 -1.35511174  -0.7551117
#> 8  08 -0.55595581 -0.55595581 -0.45595581 -0.25595581   0.3440442
#> 9  09 -0.21609756 -0.21609756 -0.11609756  0.08390244   0.6839024
#> 10 10 -1.27043591 -1.27043591 -1.17043591 -0.97043591  -0.3704359