Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.4272943 -0.4272943 -0.3272943
#> 2  02 -1.1175271 -1.1175271 -1.0175271
#> 3  03 -0.8376635 -0.8376635 -0.7376635
#> 4  04  1.1445325  1.1445325  1.2445325
#> 5  05  1.4325251  1.4325251  1.5325251
#> 6  06  0.9162028  0.9162028  1.0162028
#> 7  07  0.5166365  0.5166365  0.6166365
#> 8  08  0.2227731  0.2227731  0.3227731
#> 9  09 -0.8524291 -0.8524291 -0.7524291
#> 10 10 -0.5465845 -0.5465845 -0.4465845

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.4233494 -0.4233494 -0.32334940
#> 2  02  0.8062977  0.8062977  0.90629774
#> 3  03  1.1177047  1.1177047  1.21770466
#> 4  04  0.3554537  0.3554537  0.45545372
#> 5  05 -0.7981638 -0.7981638 -0.69816380
#> 6  06  0.2776875  0.2776875  0.37768751
#> 7  07 -0.1431655 -0.1431655 -0.04316552
#> 8  08 -1.1990663 -1.1990663 -1.09906635
#> 9  09 -2.5773505 -2.5773505 -2.47735047
#> 10 10  0.1248286  0.1248286  0.22482858

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -1.44597908 -1.34597908 -1.24597908 -1.1459791
#> 2  02  0.28250374  0.38250374  0.48250374  0.5825037
#> 3  03  0.93336739  1.03336739  1.13336739  1.2333674
#> 4  04 -0.72681518 -0.62681518 -0.52681518 -0.4268152
#> 5  05  0.05692087  0.15692087  0.25692087  0.3569209
#> 6  06 -1.21265838 -1.11265838 -1.01265838 -0.9126584
#> 7  07 -0.50510435 -0.40510435 -0.30510435 -0.2051043
#> 8  08 -0.12479468 -0.02479468  0.07520532  0.1752053
#> 9  09  1.29753749  1.39753749  1.49753749  1.5975375
#> 10 10  0.65352079  0.75352079  0.85352079  0.9535208

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.78976505  1.78976505  1.88976505  2.08976505  2.68976505
#> 2  02 -0.96523169 -0.96523169 -0.86523169 -0.66523169 -0.06523169
#> 3  03  0.94380274  0.94380274  1.04380274  1.24380274  1.84380274
#> 4  04 -0.53859145 -0.53859145 -0.43859145 -0.23859145  0.36140855
#> 5  05  0.78144947  0.78144947  0.88144947  1.08144947  1.68144947
#> 6  06  0.67233759  0.67233759  0.77233759  0.97233759  1.57233759
#> 7  07 -0.04085456 -0.04085456  0.05914544  0.25914544  0.85914544
#> 8  08 -0.36779566 -0.36779566 -0.26779566 -0.06779566  0.53220434
#> 9  09 -1.67521230 -1.67521230 -1.57521230 -1.37521230 -0.77521230
#> 10 10  0.11451499  0.11451499  0.21451499  0.41451499  1.01451499