Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  2.15425141  2.15425141  2.25425141
#> 2  02 -0.65066321 -0.65066321 -0.55066321
#> 3  03 -1.14122444 -1.14122444 -1.04122444
#> 4  04 -0.02474833 -0.02474833  0.07525167
#> 5  05 -1.18783135 -1.18783135 -1.08783135
#> 6  06  0.34768900  0.34768900  0.44768900
#> 7  07  1.03940501  1.03940501  1.13940501
#> 8  08 -1.01239475 -1.01239475 -0.91239475
#> 9  09 -1.38571605 -1.38571605 -1.28571605
#> 10 10  1.47098612  1.47098612  1.57098612

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.43321717  0.43321717  0.5332172
#> 2  02 -0.29866705 -0.29866705 -0.1986670
#> 3  03  0.30393067  0.30393067  0.4039307
#> 4  04  0.40922020  0.40922020  0.5092202
#> 5  05  1.21897763  1.21897763  1.3189776
#> 6  06  0.12678436  0.12678436  0.2267844
#> 7  07  0.06308573  0.06308573  0.1630857
#> 8  08  1.95246070  1.95246070  2.0524607
#> 9  09  1.82379037  1.82379037  1.9237904
#> 10 10 -0.39198041 -0.39198041 -0.2919804

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  1.3154410  1.4154410  1.5154410  1.6154410
#> 2  02 -1.2684596 -1.1684596 -1.0684596 -0.9684596
#> 3  03  1.0757831  1.1757831  1.2757831  1.3757831
#> 4  04 -0.5992689 -0.4992689 -0.3992689 -0.2992689
#> 5  05 -1.0220819 -0.9220819 -0.8220819 -0.7220819
#> 6  06 -2.0328566 -1.9328566 -1.8328566 -1.7328566
#> 7  07  0.2261866  0.3261866  0.4261866  0.5261866
#> 8  08  0.6518230  0.7518230  0.8518230  0.9518230
#> 9  09  0.6572215  0.7572215  0.8572215  0.9572215
#> 10 10  0.6190955  0.7190955  0.8190955  0.9190955

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.6817956  -0.6817956  -0.5817956  -0.3817956   0.2182044
#> 2  02  0.7179889   0.7179889   0.8179889   1.0179889   1.6179889
#> 3  03  0.7664134   0.7664134   0.8664134   1.0664134   1.6664134
#> 4  04  0.7524992   0.7524992   0.8524992   1.0524992   1.6524992
#> 5  05  1.7072845   1.7072845   1.8072845   2.0072845   2.6072845
#> 6  06  0.1381451   0.1381451   0.2381451   0.4381451   1.0381451
#> 7  07  0.4586931   0.4586931   0.5586931   0.7586931   1.3586931
#> 8  08 -0.4667141  -0.4667141  -0.3667141  -0.1667141   0.4332859
#> 9  09  0.2391310   0.2391310   0.3391310   0.5391310   1.1391310
#> 10 10  0.2627956   0.2627956   0.3627956   0.5627956   1.1627956