Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.7848570  0.7848570  0.88485699
#> 2  02  1.0395961  1.0395961  1.13959608
#> 3  03 -0.1227534 -0.1227534 -0.02275337
#> 4  04  1.7264902  1.7264902  1.82649020
#> 5  05  2.7458448  2.7458448  2.84584483
#> 6  06 -1.1900947 -1.1900947 -1.09009468
#> 7  07 -2.4172512 -2.4172512 -2.31725124
#> 8  08  1.0452065  1.0452065  1.14520654
#> 9  09  0.6178690  0.6178690  0.71786902
#> 10 10 -1.1744426 -1.1744426 -1.07444262

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.60262559 -0.60262559 -0.50262559
#> 2  02 -0.14139449 -0.14139449 -0.04139449
#> 3  03 -1.75847712 -1.75847712 -1.65847712
#> 4  04 -1.34640139 -1.34640139 -1.24640139
#> 5  05 -2.11547261 -2.11547261 -2.01547261
#> 6  06 -0.20110112 -0.20110112 -0.10110112
#> 7  07 -1.00947780 -1.00947780 -0.90947780
#> 8  08  0.08382126  0.08382126  0.18382126
#> 9  09 -0.56933903 -0.56933903 -0.46933903
#> 10 10 -1.14298395 -1.14298395 -1.04298395

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  1.5191103  1.61911033  1.71911033  1.8191103
#> 2  02  0.1063672  0.20636716  0.30636716  0.4063672
#> 3  03 -0.5401205 -0.44012048 -0.34012048 -0.2401205
#> 4  04 -0.4638308 -0.36383081 -0.26383081 -0.1638308
#> 5  05  0.8783904  0.97839045  1.07839045  1.1783904
#> 6  06 -0.1691586 -0.06915861  0.03084139  0.1308414
#> 7  07 -0.8788685 -0.77886845 -0.67886845 -0.5788685
#> 8  08 -1.8441609 -1.74416091 -1.64416091 -1.5441609
#> 9  09 -0.7103994 -0.61039940 -0.51039940 -0.4103994
#> 10 10  0.7599235  0.85992353  0.95992353  1.0599235

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.00906884 -0.00906884  0.09093116  0.29093116   0.8909312
#> 2  02 -1.35999553 -1.35999553 -1.25999553 -1.05999553  -0.4599955
#> 3  03 -0.32191902 -0.32191902 -0.22191902 -0.02191902   0.5780810
#> 4  04 -1.03356174 -1.03356174 -0.93356174 -0.73356174  -0.1335617
#> 5  05  0.46227895  0.46227895  0.56227895  0.76227895   1.3622789
#> 6  06  1.22188149  1.22188149  1.32188149  1.52188149   2.1218815
#> 7  07 -0.49265388 -0.49265388 -0.39265388 -0.19265388   0.4073461
#> 8  08 -0.32773343 -0.32773343 -0.22773343 -0.02773343   0.5722666
#> 9  09  2.24102124  2.24102124  2.34102124  2.54102124   3.1410212
#> 10 10 -0.42600737 -0.42600737 -0.32600737 -0.12600737   0.4739926