Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.97736423  0.97736423  1.07736423
#> 2  02  0.24642007  0.24642007  0.34642007
#> 3  03 -0.15370287 -0.15370287 -0.05370287
#> 4  04  0.46502962  0.46502962  0.56502962
#> 5  05  0.11306083  0.11306083  0.21306083
#> 6  06 -1.23997439 -1.23997439 -1.13997439
#> 7  07  0.02828342  0.02828342  0.12828342
#> 8  08  0.14938126  0.14938126  0.24938126
#> 9  09 -0.89702496 -0.89702496 -0.79702496
#> 10 10 -1.09256778 -1.09256778 -0.99256778

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.3516743  1.3516743  1.45167428
#> 2  02  0.0458090  0.0458090  0.14580900
#> 3  03  1.1895501  1.1895501  1.28955006
#> 4  04  0.2319293  0.2319293  0.33192934
#> 5  05  0.6742349  0.6742349  0.77423491
#> 6  06  1.1300075  1.1300075  1.23000750
#> 7  07 -0.1934157 -0.1934157 -0.09341572
#> 8  08 -1.2095326 -1.2095326 -1.10953259
#> 9  09  0.1914935  0.1914935  0.29149353
#> 10 10  0.1602351  0.1602351  0.26023510

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  2.0886236  2.18862360  2.28862360  2.3886236
#> 2  02 -0.5176716 -0.41767156 -0.31767156 -0.2176716
#> 3  03 -1.4636702 -1.36367022 -1.26367022 -1.1636702
#> 4  04 -2.4281577 -2.32815774 -2.22815774 -2.1281577
#> 5  05 -0.9357699 -0.83576989 -0.73576989 -0.6357699
#> 6  06 -1.3529890 -1.25298904 -1.15298904 -1.0529890
#> 7  07  0.6529038  0.75290379  0.85290379  0.9529038
#> 8  08  1.4147760  1.51477604  1.61477604  1.7147760
#> 9  09 -0.1853143 -0.08531433  0.01468567  0.1146857
#> 10 10  0.6217021  0.72170211  0.82170211  0.9217021

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.2088601  -0.2088601 -0.10886005  0.09113995  0.69113995
#> 2  02 -0.3321601  -0.3321601 -0.23216013 -0.03216013  0.56783987
#> 3  03 -0.8157161  -0.8157161 -0.71571615 -0.51571615  0.08428385
#> 4  04  0.4998876   0.4998876  0.59988765  0.79988765  1.39988765
#> 5  05  1.0785833   1.0785833  1.17858331  1.37858331  1.97858331
#> 6  06 -1.4585779  -1.4585779 -1.35857793 -1.15857793 -0.55857793
#> 7  07 -0.1717431  -0.1717431 -0.07174313  0.12825687  0.72825687
#> 8  08  0.1604886   0.1604886  0.26048863  0.46048863  1.06048863
#> 9  09  0.9392812   0.9392812  1.03928125  1.23928125  1.83928125
#> 10 10 -0.6442286  -0.6442286 -0.54422863 -0.34422863  0.25577137