Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.0926452  0.0926452  0.1926452
#> 2  02 -0.3267357 -0.3267357 -0.2267357
#> 3  03  0.5747223  0.5747223  0.6747223
#> 4  04 -0.3982541 -0.3982541 -0.2982541
#> 5  05  0.8320746  0.8320746  0.9320746
#> 6  06  0.3725336  0.3725336  0.4725336
#> 7  07 -0.2486815 -0.2486815 -0.1486815
#> 8  08 -0.1605573 -0.1605573 -0.0605573
#> 9  09  0.8846567  0.8846567  0.9846567
#> 10 10  1.6613769  1.6613769  1.7613769

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -2.1631218 -2.1631218 -2.0631218
#> 2  02  0.9029484  0.9029484  1.0029484
#> 3  03  0.5738735  0.5738735  0.6738735
#> 4  04  1.2357114  1.2357114  1.3357114
#> 5  05  2.1471986  2.1471986  2.2471986
#> 6  06  0.6155249  0.6155249  0.7155249
#> 7  07 -0.9087384 -0.9087384 -0.8087384
#> 8  08  0.6425643  0.6425643  0.7425643
#> 9  09  1.2810780  1.2810780  1.3810780
#> 10 10 -1.5172470 -1.5172470 -1.4172470

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.49475595 -0.39475595 -0.2947559 -0.1947559
#> 2  02  0.57740601  0.67740601  0.7774060  0.8774060
#> 3  03 -0.04148290  0.05851710  0.1585171  0.2585171
#> 4  04 -0.46864777 -0.36864777 -0.2686478 -0.1686478
#> 5  05  0.10033562  0.20033562  0.3003356  0.4003356
#> 6  06  1.02009671  1.12009671  1.2200967  1.3200967
#> 7  07  0.78310971  0.88310971  0.9831097  1.0831097
#> 8  08 -0.87233076 -0.77233076 -0.6723308 -0.5723308
#> 9  09 -0.06008258  0.03991742  0.1399174  0.2399174
#> 10 10  1.48919669  1.58919669  1.6891967  1.7891967

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.2084029  -1.2084029  -1.1084029  -0.9084029 -0.30840293
#> 2  02 -0.5536668  -0.5536668  -0.4536668  -0.2536668  0.34633324
#> 3  03  0.3836862   0.3836862   0.4836862   0.6836862  1.28368621
#> 4  04 -0.9592698  -0.9592698  -0.8592698  -0.6592698 -0.05926977
#> 5  05  0.5422546   0.5422546   0.6422546   0.8422546  1.44225462
#> 6  06  1.2181356   1.2181356   1.3181356   1.5181356  2.11813562
#> 7  07  0.5011164   0.5011164   0.6011164   0.8011164  1.40111642
#> 8  08 -0.4849121  -0.4849121  -0.3849121  -0.1849121  0.41508789
#> 9  09  1.3788893   1.3788893   1.4788893   1.6788893  2.27888935
#> 10 10 -0.9333119  -0.9333119  -0.8333119  -0.6333119 -0.03331188