Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.52537579  0.52537579  0.6253758
#> 2  02 -1.60374581 -1.60374581 -1.5037458
#> 3  03 -0.54581156 -0.54581156 -0.4458116
#> 4  04  0.38853192  0.38853192  0.4885319
#> 5  05  1.19922627  1.19922627  1.2992263
#> 6  06  0.39901669  0.39901669  0.4990167
#> 7  07  0.08404577  0.08404577  0.1840458
#> 8  08  1.51361230  1.51361230  1.6136123
#> 9  09  1.59524220  1.59524220  1.6952422
#> 10 10 -1.48646595 -1.48646595 -1.3864659

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.53848685 -1.53848685 -1.43848685
#> 2  02 -1.45429731 -1.45429731 -1.35429731
#> 3  03 -2.39368230 -2.39368230 -2.29368230
#> 4  04  1.09080308  1.09080308  1.19080308
#> 5  05 -0.46050230 -0.46050230 -0.36050230
#> 6  06 -0.31916749 -0.31916749 -0.21916749
#> 7  07 -1.59310123 -1.59310123 -1.49310123
#> 8  08 -0.28958439 -0.28958439 -0.18958439
#> 9  09 -0.02416231 -0.02416231  0.07583769
#> 10 10  0.20603965  0.20603965  0.30603965

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2        Y_Z_3
#> 1  01  0.89542141  0.9954214  1.0954214  1.195421410
#> 2  02  0.96073696  1.0607370  1.1607370  1.260736959
#> 3  03 -0.30236694 -0.2023669 -0.1023669 -0.002366936
#> 4  04 -0.80079794 -0.7007979 -0.6007979 -0.500797945
#> 5  05  0.09314011  0.1931401  0.2931401  0.393140113
#> 6  06  0.77715748  0.8771575  0.9771575  1.077157482
#> 7  07  1.04099874  1.1409987  1.2409987  1.340998744
#> 8  08 -1.47771309 -1.3777131 -1.2777131 -1.177713088
#> 9  09  0.15700994  0.2570099  0.3570099  0.457009939
#> 10 10  0.75835193  0.8583519  0.9583519  1.058351928

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.4815615  -0.4815615 -0.38156151 -0.18156151   0.4184385
#> 2  02 -1.9396616  -1.9396616 -1.83966156 -1.63966156  -1.0396616
#> 3  03  0.6899805   0.6899805  0.78998054  0.98998054   1.5899805
#> 4  04  1.5071146   1.5071146  1.60711463  1.80711463   2.4071146
#> 5  05 -1.0810688  -1.0810688 -0.98106884 -0.78106884  -0.1810688
#> 6  06 -0.2851413  -0.2851413 -0.18514126  0.01485874   0.6148587
#> 7  07 -0.1109174  -0.1109174 -0.01091739  0.18908261   0.7890826
#> 8  08 -1.1651110  -1.1651110 -1.06511097 -0.86511097  -0.2651110
#> 9  09 -1.7085840  -1.7085840 -1.60858400 -1.40858400  -0.8085840
#> 10 10  0.1656782   0.1656782  0.26567823  0.46567823   1.0656782