Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.01678711  0.01678711  0.1167871
#> 2  02 -0.59321362 -0.59321362 -0.4932136
#> 3  03  1.51337477  1.51337477  1.6133748
#> 4  04 -0.30775233 -0.30775233 -0.2077523
#> 5  05 -1.78764436 -1.78764436 -1.6876444
#> 6  06 -0.79220589 -0.79220589 -0.6922059
#> 7  07 -1.10622494 -1.10622494 -1.0062249
#> 8  08 -0.47471127 -0.47471127 -0.3747113
#> 9  09  0.18866134  0.18866134  0.2886613
#> 10 10 -1.04937460 -1.04937460 -0.9493746

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.20008595 -0.20008595 -0.10008595
#> 2  02 -0.29223806 -0.29223806 -0.19223806
#> 3  03 -1.29218996 -1.29218996 -1.19218996
#> 4  04 -0.84504877 -0.84504877 -0.74504877
#> 5  05 -0.08182201 -0.08182201  0.01817799
#> 6  06 -1.36637989 -1.36637989 -1.26637989
#> 7  07  0.18076961  0.18076961  0.28076961
#> 8  08 -0.64653332 -0.64653332 -0.54653332
#> 9  09  1.13494580  1.13494580  1.23494580
#> 10 10 -0.84144433 -0.84144433 -0.74144433

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.3286130 -1.2286130 -1.1286130 -1.0286130
#> 2  02 -0.6223387 -0.5223387 -0.4223387 -0.3223387
#> 3  03 -0.5858417 -0.4858417 -0.3858417 -0.2858417
#> 4  04  0.7812434  0.8812434  0.9812434  1.0812434
#> 5  05 -2.4330168 -2.3330168 -2.2330168 -2.1330168
#> 6  06  0.7201332  0.8201332  0.9201332  1.0201332
#> 7  07  0.5345124  0.6345124  0.7345124  0.8345124
#> 8  08 -2.5721580 -2.4721580 -2.3721580 -2.2721580
#> 9  09 -1.5476196 -1.4476196 -1.3476196 -1.2476196
#> 10 10  1.3004781  1.4004781  1.5004781  1.6004781

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.04329848  1.04329848   1.1432985   1.3432985  1.94329848
#> 2  02 -1.31433364 -1.31433364  -1.2143336  -1.0143336 -0.41433364
#> 3  03 -0.68828619 -0.68828619  -0.5882862  -0.3882862  0.21171381
#> 4  04 -0.76696593 -0.76696593  -0.6669659  -0.4669659  0.13303407
#> 5  05  1.28866338  1.28866338   1.3886634   1.5886634  2.18866338
#> 6  06 -0.85827594 -0.85827594  -0.7582759  -0.5582759  0.04172406
#> 7  07  0.88080214  0.88080214   0.9808021   1.1808021  1.78080214
#> 8  08  0.05113078  0.05113078   0.1511308   0.3511308  0.95113078
#> 9  09 -0.63430927 -0.63430927  -0.5343093  -0.3343093  0.26569073
#> 10 10 -1.53775207 -1.53775207  -1.4377521  -1.2377521 -0.63775207