Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.05799538 -0.05799538  0.04200462
#> 2  02 -0.86686857 -0.86686857 -0.76686857
#> 3  03 -1.04391377 -1.04391377 -0.94391377
#> 4  04  1.29631631  1.29631631  1.39631631
#> 5  05  0.46592670  0.46592670  0.56592670
#> 6  06 -1.15107003 -1.15107003 -1.05107003
#> 7  07  0.04314544  0.04314544  0.14314544
#> 8  08  1.02794426  1.02794426  1.12794426
#> 9  09  0.58459111  0.58459111  0.68459111
#> 10 10  0.20994238  0.20994238  0.30994238

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01 -0.17272129 -0.17272129 -0.072721286
#> 2  02 -0.32023557 -0.32023557 -0.220235573
#> 3  03  1.36862363  1.36862363  1.468623634
#> 4  04 -1.07933683 -1.07933683 -0.979336829
#> 5  05 -0.10251791 -0.10251791 -0.002517907
#> 6  06  0.09127899  0.09127899  0.191278989
#> 7  07 -1.33211605 -1.33211605 -1.232116052
#> 8  08  0.05264606  0.05264606  0.152646057
#> 9  09 -1.39506381 -1.39506381 -1.295063814
#> 10 10  2.39159976  2.39159976  2.491599760

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  0.9714841  1.07148412  1.17148412  1.2714841
#> 2  02 -1.5521207 -1.45212066 -1.35212066 -1.2521207
#> 3  03  1.8319420  1.93194195  2.03194195  2.1319420
#> 4  04 -1.4108253 -1.31082526 -1.21082526 -1.1108253
#> 5  05 -2.2331079 -2.13310788 -2.03310788 -1.9331079
#> 6  06 -0.1506709 -0.05067094  0.04932906  0.1493291
#> 7  07 -0.1315710 -0.03157100  0.06842900  0.1684290
#> 8  08  0.2836838  0.38368381  0.48368381  0.5836838
#> 9  09  0.9730980  1.07309803  1.17309803  1.2730980
#> 10 10 -1.4831018 -1.38310177 -1.28310177 -1.1831018

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5689518   0.5689518  0.66895177  0.868951768   1.4689518
#> 2  02  0.8085231   0.8085231  0.90852311  1.108523114   1.7085231
#> 3  03  0.2698595   0.2698595  0.36985951  0.569859507   1.1698595
#> 4  04  0.8565157   0.8565157  0.95651574  1.156515742   1.7565157
#> 5  05  0.2869439   0.2869439  0.38694388  0.586943881   1.1869439
#> 6  06  1.8047815   1.8047815  1.90478146  2.104781462   2.7047815
#> 7  07 -0.2905159  -0.2905159 -0.19051591  0.009484092   0.6094841
#> 8  08 -0.1823783  -0.1823783 -0.08237832  0.117621675   0.7176217
#> 9  09 -1.4363506  -1.4363506 -1.33635059 -1.136350595  -0.5363506
#> 10 10  0.5029931   0.5029931  0.60299309  0.802993094   1.4029931