Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.1562576  0.1562576  0.2562576
#> 2  02 -0.4400731 -0.4400731 -0.3400731
#> 3  03 -0.6253134 -0.6253134 -0.5253134
#> 4  04 -0.2667195 -0.2667195 -0.1667195
#> 5  05 -0.9787651 -0.9787651 -0.8787651
#> 6  06 -0.2608765 -0.2608765 -0.1608765
#> 7  07 -1.0049529 -1.0049529 -0.9049529
#> 8  08  1.8238129  1.8238129  1.9238129
#> 9  09 -0.9389800 -0.9389800 -0.8389800
#> 10 10  0.5889988  0.5889988  0.6889988

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1017774 -1.1017774 -1.0017774
#> 2  02 -1.4587044 -1.4587044 -1.3587044
#> 3  03  0.8874626  0.8874626  0.9874626
#> 4  04  0.2653609  0.2653609  0.3653609
#> 5  05  0.8787632  0.8787632  0.9787632
#> 6  06  0.8038365  0.8038365  0.9038365
#> 7  07  0.5101560  0.5101560  0.6101560
#> 8  08  0.6677853  0.6677853  0.7677853
#> 9  09  0.3016883  0.3016883  0.4016883
#> 10 10 -1.3477909 -1.3477909 -1.2477909

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.06301562  0.163015622  0.2630156  0.3630156
#> 2  02 -0.09284321  0.007156786  0.1071568  0.2071568
#> 3  03 -0.57960177 -0.479601775 -0.3796018 -0.2796018
#> 4  04  1.61532397  1.715323966  1.8153240  1.9153240
#> 5  05  1.13174639  1.231746393  1.3317464  1.4317464
#> 6  06 -0.09778584  0.002214159  0.1022142  0.2022142
#> 7  07  0.95522783  1.055227832  1.1552278  1.2552278
#> 8  08  0.11480491  0.214804911  0.3148049  0.4148049
#> 9  09 -1.27637934 -1.176379345 -1.0763793 -0.9763793
#> 10 10 -3.02721109 -2.927211092 -2.8272111 -2.7272111

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.50497631 -0.50497631  -0.4049763  -0.2049763   0.3950237
#> 2  02 -0.66267267 -0.66267267  -0.5626727  -0.3626727   0.2373273
#> 3  03 -0.75856446 -0.75856446  -0.6585645  -0.4585645   0.1414355
#> 4  04  0.19677988  0.19677988   0.2967799   0.4967799   1.0967799
#> 5  05  1.49381173  1.49381173   1.5938117   1.7938117   2.3938117
#> 6  06  0.08782298  0.08782298   0.1878230   0.3878230   0.9878230
#> 7  07 -0.51305932 -0.51305932  -0.4130593  -0.2130593   0.3869407
#> 8  08  1.41465639  1.41465639   1.5146564   1.7146564   2.3146564
#> 9  09  2.07164046  2.07164046   2.1716405   2.3716405   2.9716405
#> 10 10  1.25366327  1.25366327   1.3536633   1.5536633   2.1536633