Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01  0.290122019  0.290122019  0.3901220
#> 2  02  0.003239148  0.003239148  0.1032391
#> 3  03  0.918923825  0.918923825  1.0189238
#> 4  04 -1.364184463 -1.364184463 -1.2641845
#> 5  05  0.811380431  0.811380431  0.9113804
#> 6  06 -0.379697332 -0.379697332 -0.2796973
#> 7  07  0.752153036  0.752153036  0.8521530
#> 8  08 -0.824650545 -0.824650545 -0.7246505
#> 9  09  0.200314224  0.200314224  0.3003142
#> 10 10  1.004887407  1.004887407  1.1048874

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.7513122 -0.7513122 -0.65131221
#> 2  02 -1.1591591 -1.1591591 -1.05915905
#> 3  03 -0.6046944 -0.6046944 -0.50469438
#> 4  04 -0.1562297 -0.1562297 -0.05622967
#> 5  05 -0.9035919 -0.9035919 -0.80359190
#> 6  06  1.2204533  1.2204533  1.32045328
#> 7  07 -0.2401735 -0.2401735 -0.14017352
#> 8  08 -0.5576783 -0.5576783 -0.45767828
#> 9  09  0.9122387  0.9122387  1.01223873
#> 10 10 -0.8868238 -0.8868238 -0.78682377

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.9032383 -0.8032383 -0.70323825 -0.60323825
#> 2  02  0.5887141  0.6887141  0.78871414  0.88871414
#> 3  03 -1.2823535 -1.1823535 -1.08235351 -0.98235351
#> 4  04 -0.2624560 -0.1624560 -0.06245603  0.03754397
#> 5  05 -1.3398448 -1.2398448 -1.13984479 -1.03984479
#> 6  06 -1.4896579 -1.3896579 -1.28965795 -1.18965795
#> 7  07  1.3293088  1.4293088  1.52930883  1.62930883
#> 8  08  0.6280880  0.7280880  0.82808796  0.92808796
#> 9  09  0.4193063  0.5193063  0.61930631  0.71930631
#> 10 10  0.2310340  0.3310340  0.43103395  0.53103395

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.15376102 -0.15376102 -0.05376102  0.14623898   0.7462390
#> 2  02 -1.52498026 -1.52498026 -1.42498026 -1.22498026  -0.6249803
#> 3  03 -0.28270862 -0.28270862 -0.18270862  0.01729138   0.6172914
#> 4  04 -0.48672320 -0.48672320 -0.38672320 -0.18672320   0.4132768
#> 5  05 -0.24892336 -0.24892336 -0.14892336  0.05107664   0.6510766
#> 6  06  0.18156769  0.18156769  0.28156769  0.48156769   1.0815677
#> 7  07  0.02206853  0.02206853  0.12206853  0.32206853   0.9220685
#> 8  08  0.98607169  0.98607169  1.08607169  1.28607169   1.8860717
#> 9  09 -1.61100017 -1.61100017 -1.51100017 -1.31100017  -0.7110002
#> 10 10 -0.53400151 -0.53400151 -0.43400151 -0.23400151   0.3659985