Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1569147 -1.1569147 -1.0569147
#> 2  02  0.5131599  0.5131599  0.6131599
#> 3  03 -0.3718664 -0.3718664 -0.2718664
#> 4  04 -0.7234422 -0.7234422 -0.6234422
#> 5  05 -0.2156155 -0.2156155 -0.1156155
#> 6  06 -1.0784120 -1.0784120 -0.9784120
#> 7  07 -0.6029350 -0.6029350 -0.5029350
#> 8  08  0.9876673  0.9876673  1.0876673
#> 9  09  0.3335000  0.3335000  0.4335000
#> 10 10  0.4045050  0.4045050  0.5045050

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1611291 -1.1611291 -1.0611291
#> 2  02  0.1516456  0.1516456  0.2516456
#> 3  03  0.5710403  0.5710403  0.6710403
#> 4  04  0.4957053  0.4957053  0.5957053
#> 5  05  1.4784436  1.4784436  1.5784436
#> 6  06  2.9432822  2.9432822  3.0432822
#> 7  07  0.4687726  0.4687726  0.5687726
#> 8  08  1.3402299  1.3402299  1.4402299
#> 9  09 -0.8873295 -0.8873295 -0.7873295
#> 10 10  1.6907529  1.6907529  1.7907529

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  1.52943946  1.62943946  1.7294395  1.82943946
#> 2  02  0.20520502  0.30520502  0.4052050  0.50520502
#> 3  03  0.41856832  0.51856832  0.6185683  0.71856832
#> 4  04 -0.43807175 -0.33807175 -0.2380718 -0.13807175
#> 5  05  0.18625039  0.28625039  0.3862504  0.48625039
#> 6  06 -0.07906347  0.02093653  0.1209365  0.22093653
#> 7  07  0.15257665  0.25257665  0.3525767  0.45257665
#> 8  08 -0.37641374 -0.27641374 -0.1764137 -0.07641374
#> 9  09 -1.54982350 -1.44982350 -1.3498235 -1.24982350
#> 10 10  1.90656328  2.00656328  2.1065633  2.20656328

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.92707202  0.92707202  1.02707202   1.2270720   1.8270720
#> 2  02 -0.19384594 -0.19384594 -0.09384594   0.1061541   0.7061541
#> 3  03 -1.07390862 -1.07390862 -0.97390862  -0.7739086  -0.1739086
#> 4  04 -1.49940983 -1.49940983 -1.39940983  -1.1994098  -0.5994098
#> 5  05  0.08819758  0.08819758  0.18819758   0.3881976   0.9881976
#> 6  06 -2.79478182 -2.79478182 -2.69478182  -2.4947818  -1.8947818
#> 7  07 -0.58019609 -0.58019609 -0.48019609  -0.2801961   0.3198039
#> 8  08  1.94725610  1.94725610  2.04725610   2.2472561   2.8472561
#> 9  09  0.09147560  0.09147560  0.19147560   0.3914756   0.9914756
#> 10 10  1.62211603  1.62211603  1.72211603   1.9221160   2.5221160