Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0        Y_Z_1
#> 1  01 -0.1073973 -0.1073973 -0.007397311
#> 2  02  0.8781050  0.8781050  0.978104955
#> 3  03  2.6554876  2.6554876  2.755487650
#> 4  04  1.3434718  1.3434718  1.443471810
#> 5  05 -2.6262218 -2.6262218 -2.526221787
#> 6  06 -0.1246690 -0.1246690 -0.024669029
#> 7  07 -1.8894169 -1.8894169 -1.789416895
#> 8  08  0.7070309  0.7070309  0.807030906
#> 9  09  0.1882986  0.1882986  0.288298568
#> 10 10  0.3502163  0.3502163  0.450216267

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.1182016  0.1182016  0.21820160
#> 2  02 -0.8930003 -0.8930003 -0.79300027
#> 3  03  0.1803642  0.1803642  0.28036424
#> 4  04 -0.2800260 -0.2800260 -0.18002601
#> 5  05 -0.1871727 -0.1871727 -0.08717268
#> 6  06  0.4403874  0.4403874  0.54038736
#> 7  07 -1.0654603 -1.0654603 -0.96546025
#> 8  08 -0.4901072 -0.4901072 -0.39010724
#> 9  09 -0.4666386 -0.4666386 -0.36663862
#> 10 10 -1.4211180 -1.4211180 -1.32111796

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -1.1365405 -1.0365405 -0.93654052 -0.83654052
#> 2  02  0.6143298  0.7143298  0.81432981  0.91432981
#> 3  03 -0.2434272 -0.1434272 -0.04342719  0.05657281
#> 4  04 -1.9867490 -1.8867490 -1.78674899 -1.68674899
#> 5  05  0.2111301  0.3111301  0.41113008  0.51113008
#> 6  06  1.1527730  1.2527730  1.35277304  1.45277304
#> 7  07  1.0388426  1.1388426  1.23884255  1.33884255
#> 8  08 -1.4493165 -1.3493165 -1.24931653 -1.14931653
#> 9  09  0.2912295  0.3912295  0.49122945  0.59122945
#> 10 10 -0.2069152 -0.1069152 -0.00691522  0.09308478

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1  Y_Z1_1_Z2_1
#> 1  01  0.429338559  0.429338559  0.52933856  0.729338559  1.329338559
#> 2  02  0.426351557  0.426351557  0.52635156  0.726351557  1.326351557
#> 3  03 -0.009376579 -0.009376579  0.09062342  0.290623421  0.890623421
#> 4  04  0.903740742  0.903740742  1.00374074  1.203740742  1.803740742
#> 5  05 -1.187722982 -1.187722982 -1.08772298 -0.887722982 -0.287722982
#> 6  06  0.443407623  0.443407623  0.54340762  0.743407623  1.343407623
#> 7  07 -0.295086784 -0.295086784 -0.19508678  0.004913216  0.604913216
#> 8  08 -2.166080856 -2.166080856 -2.06608086 -1.866080856 -1.266080856
#> 9  09  0.965514317  0.965514317  1.06551432  1.265514317  1.865514317
#> 10 10 -0.894809086 -0.894809086 -0.79480909 -0.594809086  0.005190914