Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8493748 -0.8493748 -0.7493748
#> 2  02  0.6677181  0.6677181  0.7677181
#> 3  03  0.7199600  0.7199600  0.8199600
#> 4  04  0.7061508  0.7061508  0.8061508
#> 5  05  1.2759913  1.2759913  1.3759913
#> 6  06 -0.5723955 -0.5723955 -0.4723955
#> 7  07 -1.1913787 -1.1913787 -1.0913787
#> 8  08  0.3063256  0.3063256  0.4063256
#> 9  09 -0.2879729 -0.2879729 -0.1879729
#> 10 10  0.2038680  0.2038680  0.3038680

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.1505906  0.1505906  0.2505906
#> 2  02  0.6122534  0.6122534  0.7122534
#> 3  03 -0.5049485 -0.5049485 -0.4049485
#> 4  04 -1.5245937 -1.5245937 -1.4245937
#> 5  05 -0.0181122 -0.0181122  0.0818878
#> 6  06 -0.8733214 -0.8733214 -0.7733214
#> 7  07  1.8219211  1.8219211  1.9219211
#> 8  08  0.3153429  0.3153429  0.4153429
#> 9  09  2.3507865  2.3507865  2.4507865
#> 10 10  0.3150640  0.3150640  0.4150640

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.547940040 -0.4479400 -0.3479400 -0.2479400
#> 2  02 -2.805728575 -2.7057286 -2.6057286 -2.5057286
#> 3  03 -0.778433031 -0.6784330 -0.5784330 -0.4784330
#> 4  04  0.007355818  0.1073558  0.2073558  0.3073558
#> 5  05  0.612627998  0.7126280  0.8126280  0.9126280
#> 6  06  0.907582398  1.0075824  1.1075824  1.2075824
#> 7  07  0.091274901  0.1912749  0.2912749  0.3912749
#> 8  08  0.739144319  0.8391443  0.9391443  1.0391443
#> 9  09 -2.120233648 -2.0202336 -1.9202336 -1.8202336
#> 10 10  0.382726002  0.4827260  0.5827260  0.6827260

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.30337891  0.30337891   0.4033789  0.60337891  1.20337891
#> 2  02 -0.51762826 -0.51762826  -0.4176283 -0.21762826  0.38237174
#> 3  03  0.28097331  0.28097331   0.3809733  0.58097331  1.18097331
#> 4  04 -0.68318557 -0.68318557  -0.5831856 -0.38318557  0.21681443
#> 5  05  0.08306091  0.08306091   0.1830609  0.38306091  0.98306091
#> 6  06 -0.32502934 -0.32502934  -0.2250293 -0.02502934  0.57497066
#> 7  07  0.95256730  0.95256730   1.0525673  1.25256730  1.85256730
#> 8  08 -0.37472266 -0.37472266  -0.2747227 -0.07472266  0.52527734
#> 9  09 -0.99448659 -0.99448659  -0.8944866 -0.69448659 -0.09448659
#> 10 10  0.82156196  0.82156196   0.9215620  1.12156196  1.72156196