Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.83324503 -0.83324503 -0.73324503
#> 2  02  0.08108301  0.08108301  0.18108301
#> 3  03  0.36857736  0.36857736  0.46857736
#> 4  04 -1.69386462 -1.69386462 -1.59386462
#> 5  05 -1.46806927 -1.46806927 -1.36806927
#> 6  06 -0.88508835 -0.88508835 -0.78508835
#> 7  07 -0.08290015 -0.08290015  0.01709985
#> 8  08  0.24782598  0.24782598  0.34782598
#> 9  09 -0.57567763 -0.57567763 -0.47567763
#> 10 10 -1.20037372 -1.20037372 -1.10037372

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8844926 -0.8844926 -0.7844926
#> 2  02  1.1578970  1.1578970  1.2578970
#> 3  03 -1.6939844 -1.6939844 -1.5939844
#> 4  04  0.9073033  0.9073033  1.0073033
#> 5  05  1.0702794  1.0702794  1.1702794
#> 6  06 -0.5909916 -0.5909916 -0.4909916
#> 7  07  0.9017379  0.9017379  1.0017379
#> 8  08  1.1028510  1.1028510  1.2028510
#> 9  09 -1.1660884 -1.1660884 -1.0660884
#> 10 10  0.6036110  0.6036110  0.7036110

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -2.2862217 -2.1862217 -2.0862217 -1.9862217
#> 2  02  1.9422638  2.0422638  2.1422638  2.2422638
#> 3  03  0.8078536  0.9078536  1.0078536  1.1078536
#> 4  04 -1.1473691 -1.0473691 -0.9473691 -0.8473691
#> 5  05  0.7131041  0.8131041  0.9131041  1.0131041
#> 6  06  0.4442989  0.5442989  0.6442989  0.7442989
#> 7  07  0.5737139  0.6737139  0.7737139  0.8737139
#> 8  08 -1.0133682 -0.9133682 -0.8133682 -0.7133682
#> 9  09 -2.1505909 -2.0505909 -1.9505909 -1.8505909
#> 10 10 -1.6953152 -1.5953152 -1.4953152 -1.3953152

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1   Y_Z1_1_Z2_1
#> 1  01  0.3825813   0.3825813   0.4825813   0.6825813  1.2825812567
#> 2  02  0.7767574   0.7767574   0.8767574   1.0767574  1.6767573658
#> 3  03  1.0956458   1.0956458   1.1956458   1.3956458  1.9956457623
#> 4  04 -1.8595543  -1.8595543  -1.7595543  -1.5595543 -0.9595543265
#> 5  05  0.5925011   0.5925011   0.6925011   0.8925011  1.4925011196
#> 6  06  0.6927479   0.6927479   0.7927479   0.9927479  1.5927478973
#> 7  07  0.1665085   0.1665085   0.2665085   0.4665085  1.0665084642
#> 8  08 -0.8992286  -0.8992286  -0.7992286  -0.5992286  0.0007713779
#> 9  09  0.3470852   0.3470852   0.4470852   0.6470852  1.2470852444
#> 10 10 -1.6319396  -1.6319396  -1.5319396  -1.3319396 -0.7319396089