Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.42642897 -0.42642897 -0.32642897
#> 2  02 -0.27203037 -0.27203037 -0.17203037
#> 3  03 -0.89835202 -0.89835202 -0.79835202
#> 4  04 -0.01244898 -0.01244898  0.08755102
#> 5  05 -1.73391489 -1.73391489 -1.63391489
#> 6  06 -1.45301237 -1.45301237 -1.35301237
#> 7  07  0.17085781  0.17085781  0.27085781
#> 8  08  0.58008904  0.58008904  0.68008904
#> 9  09 -0.64508091 -0.64508091 -0.54508091
#> 10 10 -0.04197052 -0.04197052  0.05802948

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.5462010  0.5462010  0.6462010
#> 2  02 -1.5731429 -1.5731429 -1.4731429
#> 3  03  0.9399182  0.9399182  1.0399182
#> 4  04  0.1341124  0.1341124  0.2341124
#> 5  05 -0.4039796 -0.4039796 -0.3039796
#> 6  06  0.9170127  0.9170127  1.0170127
#> 7  07  0.5422164  0.5422164  0.6422164
#> 8  08  1.4299095  1.4299095  1.5299095
#> 9  09 -2.0574405 -2.0574405 -1.9574405
#> 10 10 -2.3764462 -2.3764462 -2.2764462

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2         Y_Z_3
#> 1  01 -1.6253980 -1.5253980 -1.4253980 -1.3253980016
#> 2  02  0.9653419  1.0653419  1.1653419  1.2653418949
#> 3  03  0.1551903  0.2551903  0.3551903  0.4551903062
#> 4  04  2.7242496  2.8242496  2.9242496  3.0242495828
#> 5  05 -0.3248281 -0.2248281 -0.1248281 -0.0248280549
#> 6  06  0.8013886  0.9013886  1.0013886  1.1013885841
#> 7  07 -0.3009669 -0.2009669 -0.1009669 -0.0009669202
#> 8  08 -0.3129301 -0.2129301 -0.1129301 -0.0129301337
#> 9  09  1.0555005  1.1555005  1.2555005  1.3555005157
#> 10 10  0.1290973  0.2290973  0.3290973  0.4290973120

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.4759017   0.4759017  0.57590167   0.7759017  1.37590167
#> 2  02 -0.8049406  -0.8049406 -0.70494064  -0.5049406  0.09505936
#> 3  03  1.1426823   1.1426823  1.24268230   1.4426823  2.04268230
#> 4  04 -0.2713281  -0.2713281 -0.17132810   0.0286719  0.62867190
#> 5  05 -1.2296961  -1.2296961 -1.12969613  -0.9296961 -0.32969613
#> 6  06 -0.1663320  -0.1663320 -0.06633197   0.1336680  0.73366803
#> 7  07 -0.6649734  -0.6649734 -0.56497337  -0.3649734  0.23502663
#> 8  08  1.3570763   1.3570763  1.45707633   1.6570763  2.25707633
#> 9  09 -0.6867812  -0.6867812 -0.58678122  -0.3867812  0.21321878
#> 10 10  0.4897959   0.4897959  0.58979590   0.7897959  1.38979590