Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.2900976  0.2900976  0.3900976
#> 2  02  0.5709383  0.5709383  0.6709383
#> 3  03 -1.6637859 -1.6637859 -1.5637859
#> 4  04 -0.4179094 -0.4179094 -0.3179094
#> 5  05 -0.3779787 -0.3779787 -0.2779787
#> 6  06 -1.6128523 -1.6128523 -1.5128523
#> 7  07  0.3652174  0.3652174  0.4652174
#> 8  08  1.0164151  1.0164151  1.1164151
#> 9  09 -0.6828916 -0.6828916 -0.5828916
#> 10 10  1.9006998  1.9006998  2.0006998

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.0908151  1.0908151  1.1908151
#> 2  02 -1.7955292 -1.7955292 -1.6955292
#> 3  03  0.1868936  0.1868936  0.2868936
#> 4  04 -1.9907489 -1.9907489 -1.8907489
#> 5  05  0.5464664  0.5464664  0.6464664
#> 6  06  1.9373618  1.9373618  2.0373618
#> 7  07  1.8656998  1.8656998  1.9656998
#> 8  08 -0.8101974 -0.8101974 -0.7101974
#> 9  09 -0.3559712 -0.3559712 -0.2559712
#> 10 10  0.5308959  0.5308959  0.6308959

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -0.34902705 -0.24902705 -0.1490270 -0.04902705
#> 2  02 -2.44361513 -2.34361513 -2.2436151 -2.14361513
#> 3  03  0.23685333  0.33685333  0.4368533  0.53685333
#> 4  04  1.30242087  1.40242087  1.5024209  1.60242087
#> 5  05  0.24458077  0.34458077  0.4445808  0.54458077
#> 6  06 -0.06137716  0.03862284  0.1386228  0.23862284
#> 7  07 -1.77976553 -1.67976553 -1.5797655 -1.47976553
#> 8  08  1.97834804  2.07834804  2.1783480  2.27834804
#> 9  09 -0.94505047 -0.84505047 -0.7450505 -0.64505047
#> 10 10  0.70083516  0.80083516  0.9008352  1.00083516

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.0298333  -1.0298333 -0.92983330 -0.729833305  -0.1298333
#> 2  02 -0.2972536  -0.2972536 -0.19725356  0.002746442   0.6027464
#> 3  03 -1.0154473  -1.0154473 -0.91544731 -0.715447307  -0.1154473
#> 4  04  1.0686427   1.0686427  1.16864271  1.368642715   1.9686427
#> 5  05 -0.6847613  -0.6847613 -0.58476130 -0.384761300   0.2152387
#> 6  06 -0.1709785  -0.1709785 -0.07097851  0.129021491   0.7290215
#> 7  07  1.0068129   1.0068129  1.10681291  1.306812914   1.9068129
#> 8  08  0.2676363   0.2676363  0.36763631  0.567636307   1.1676363
#> 9  09  0.3029060   0.3029060  0.40290600  0.602906001   1.2029060
#> 10 10 -0.3705667  -0.3705667 -0.27056668 -0.070566684   0.5294333