Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.3597887 -0.3597887 -0.25978869
#> 2  02 -1.1258801 -1.1258801 -1.02588009
#> 3  03 -0.8590574 -0.8590574 -0.75905735
#> 4  04 -1.2640865 -1.2640865 -1.16408652
#> 5  05 -0.1720313 -0.1720313 -0.07203135
#> 6  06  0.9181247  0.9181247  1.01812471
#> 7  07 -0.6567217 -0.6567217 -0.55672166
#> 8  08 -3.4250588 -3.4250588 -3.32505876
#> 9  09 -0.3012392 -0.3012392 -0.20123919
#> 10 10 -0.5137501 -0.5137501 -0.41375010

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1603076 -0.1603076 -0.06030758
#> 2  02 -0.9439339 -0.9439339 -0.84393391
#> 3  03 -1.4380429 -1.4380429 -1.33804290
#> 4  04  0.5990841  0.5990841  0.69908409
#> 5  05  1.1030943  1.1030943  1.20309431
#> 6  06 -0.8333953 -0.8333953 -0.73339527
#> 7  07 -0.1420327 -0.1420327 -0.04203265
#> 8  08  1.4935505  1.4935505  1.59355054
#> 9  09 -1.1848674 -1.1848674 -1.08486738
#> 10 10 -1.1298082 -1.1298082 -1.02980816

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.3810079  1.48100792  1.58100792  1.68100792
#> 2  02  0.7184335  0.81843354  0.91843354  1.01843354
#> 3  03  0.5613910  0.66139104  0.76139104  0.86139104
#> 4  04  0.1464364  0.24643638  0.34643638  0.44643638
#> 5  05  0.9849104  1.08491038  1.18491038  1.28491038
#> 6  06 -0.2289297 -0.12892972 -0.02892972  0.07107028
#> 7  07 -1.2712075 -1.17120753 -1.07120753 -0.97120753
#> 8  08 -1.5224618 -1.42246175 -1.32246175 -1.22246175
#> 9  09  0.9904176  1.09041764  1.19041764  1.29041764
#> 10 10 -0.1473300 -0.04732999  0.05267001  0.15267001

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.20396747  1.20396747  1.30396747  1.50396747   2.1039675
#> 2  02  1.31856209  1.31856209  1.41856209  1.61856209   2.2185621
#> 3  03  1.49475667  1.49475667  1.59475667  1.79475667   2.3947567
#> 4  04 -0.50486383 -0.50486383 -0.40486383 -0.20486383   0.3951362
#> 5  05  0.85266789  0.85266789  0.95266789  1.15266789   1.7526679
#> 6  06 -1.29201554 -1.29201554 -1.19201554 -0.99201554  -0.3920155
#> 7  07 -0.02590484 -0.02590484  0.07409516  0.27409516   0.8740952
#> 8  08 -0.23537572 -0.23537572 -0.13537572  0.06462428   0.6646243
#> 9  09 -2.83497315 -2.83497315 -2.73497315 -2.53497315  -1.9349731
#> 10 10  0.13329623  0.13329623  0.23329623  0.43329623   1.0332962