Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.3087752 -0.3087752 -0.2087752
#> 2  02 -0.4269823 -0.4269823 -0.3269823
#> 3  03 -0.8945572 -0.8945572 -0.7945572
#> 4  04 -1.6453186 -1.6453186 -1.5453186
#> 5  05 -0.8133026 -0.8133026 -0.7133026
#> 6  06 -0.2500932 -0.2500932 -0.1500932
#> 7  07  0.1587162  0.1587162  0.2587162
#> 8  08 -2.0001141 -2.0001141 -1.9001141
#> 9  09 -0.6582726 -0.6582726 -0.5582726
#> 10 10  1.6698661  1.6698661  1.7698661

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8793444 -0.8793444 -0.7793444
#> 2  02  0.9036078  0.9036078  1.0036078
#> 3  03  0.5528625  0.5528625  0.6528625
#> 4  04 -1.7395762 -1.7395762 -1.6395762
#> 5  05  0.1742763  0.1742763  0.2742763
#> 6  06 -0.6264751 -0.6264751 -0.5264751
#> 7  07  1.8174513  1.8174513  1.9174513
#> 8  08 -1.8643521 -1.8643521 -1.7643521
#> 9  09 -0.3109529 -0.3109529 -0.2109529
#> 10 10  0.1881472  0.1881472  0.2881472

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.82882026 -0.728820262 -0.62882026 -0.52882026
#> 2  02 -0.80485755 -0.704857546 -0.60485755 -0.50485755
#> 3  03 -0.90654365 -0.806543652 -0.70654365 -0.60654365
#> 4  04 -0.18776277 -0.087762767  0.01223723  0.11223723
#> 5  05 -1.63606619 -1.536066192 -1.43606619 -1.33606619
#> 6  06  1.02469917  1.124699172  1.22469917  1.32469917
#> 7  07 -0.06524723  0.034752767  0.13475277  0.23475277
#> 8  08 -0.10286353 -0.002863533  0.09713647  0.19713647
#> 9  09 -0.23947515 -0.139475146 -0.03947515  0.06052485
#> 10 10  1.12627376  1.226273760  1.32627376  1.42627376

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.8805613  -0.8805613 -0.78056131  -0.5805613  0.01943869
#> 2  02 -1.2219350  -1.2219350 -1.12193502  -0.9219350 -0.32193502
#> 3  03  0.5487934   0.5487934  0.64879341   0.8487934  1.44879341
#> 4  04 -0.9637524  -0.9637524 -0.86375236  -0.6637524 -0.06375236
#> 5  05 -0.5406336  -0.5406336 -0.44063357  -0.2406336  0.35936643
#> 6  06  1.4952491   1.4952491  1.59524915   1.7952491  2.39524915
#> 7  07 -0.1485802  -0.1485802 -0.04858016   0.1514198  0.75141984
#> 8  08  0.1098127   0.1098127  0.20981270   0.4098127  1.00981270
#> 9  09 -1.6623486  -1.6623486 -1.56234857  -1.3623486 -0.76234857
#> 10 10  1.3246687   1.3246687  1.42466869   1.6246687  2.22466869