Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.45575870  0.45575870  0.5557587
#> 2  02 -0.28811795 -0.28811795 -0.1881179
#> 3  03 -1.88396506 -1.88396506 -1.7839651
#> 4  04  0.20276880  0.20276880  0.3027688
#> 5  05 -0.47980072 -0.47980072 -0.3798007
#> 6  06  0.25112048  0.25112048  0.3511205
#> 7  07  0.01712906  0.01712906  0.1171291
#> 8  08  0.41595545  0.41595545  0.5159554
#> 9  09 -0.69992777 -0.69992777 -0.5999278
#> 10 10 -1.14294723 -1.14294723 -1.0429472

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.0373190  1.0373190  1.13731900
#> 2  02 -0.2537847 -0.2537847 -0.15378472
#> 3  03  0.0423406  0.0423406  0.14234060
#> 4  04 -2.4741384 -2.4741384 -2.37413841
#> 5  05  2.3053623  2.3053623  2.40536232
#> 6  06 -0.1191386 -0.1191386 -0.01913855
#> 7  07  0.6780719  0.6780719  0.77807192
#> 8  08 -0.3175298 -0.3175298 -0.21752983
#> 9  09 -0.1877386 -0.1877386 -0.08773860
#> 10 10  0.5564429  0.5564429  0.65644290

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.24664362  0.3466436  0.4466436  0.54664362
#> 2  02  0.04670465  0.1467046  0.2467046  0.34670465
#> 3  03 -0.54418563 -0.4441856 -0.3441856 -0.24418563
#> 4  04 -0.33948368 -0.2394837 -0.1394837 -0.03948368
#> 5  05  0.63848777  0.7384878  0.8384878  0.93848777
#> 6  06 -0.94049755 -0.8404976 -0.7404976 -0.64049755
#> 7  07  1.18763640  1.2876364  1.3876364  1.48763640
#> 8  08  1.09236430  1.1923643  1.2923643  1.39236430
#> 9  09  0.13432124  0.2343212  0.3343212  0.43432124
#> 10 10 -2.08960549 -1.9896055 -1.8896055 -1.78960549

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.84775392  0.84775392   0.9477539   1.1477539  1.74775392
#> 2  02  1.53607692  1.53607692   1.6360769   1.8360769  2.43607692
#> 3  03 -1.16025653 -1.16025653  -1.0602565  -0.8602565 -0.26025653
#> 4  04  1.29100641  1.29100641   1.3910064   1.5910064  2.19100641
#> 5  05  0.22426644  0.22426644   0.3242664   0.5242664  1.12426644
#> 6  06  0.02981518  0.02981518   0.1298152   0.3298152  0.92981518
#> 7  07  1.21692184  1.21692184   1.3169218   1.5169218  2.11692184
#> 8  08  0.60866953  0.60866953   0.7086695   0.9086695  1.50866953
#> 9  09  1.11918832  1.11918832   1.2191883   1.4191883  2.01918832
#> 10 10 -0.86063018 -0.86063018  -0.7606302  -0.5606302  0.03936982