Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.1936271  1.1936271  1.2936271
#> 2  02 -0.3254231 -0.3254231 -0.2254231
#> 3  03  1.4075608  1.4075608  1.5075608
#> 4  04  0.5319945  0.5319945  0.6319945
#> 5  05 -0.4780368 -0.4780368 -0.3780368
#> 6  06 -0.3163792 -0.3163792 -0.2163792
#> 7  07 -1.0969824 -1.0969824 -0.9969824
#> 8  08  0.4541386  0.4541386  0.5541386
#> 9  09 -1.5394812 -1.5394812 -1.4394812
#> 10 10 -0.5481319 -0.5481319 -0.4481319

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  3.2681744  3.2681744  3.3681744
#> 2  02 -1.5713962 -1.5713962 -1.4713962
#> 3  03  0.3953219  0.3953219  0.4953219
#> 4  04  0.4341700  0.4341700  0.5341700
#> 5  05 -2.3493344 -2.3493344 -2.2493344
#> 6  06  0.7934733  0.7934733  0.8934733
#> 7  07  1.4954884  1.4954884  1.5954884
#> 8  08 -0.5897527 -0.5897527 -0.4897527
#> 9  09  0.6408270  0.6408270  0.7408270
#> 10 10 -0.9060402 -0.9060402 -0.8060402

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.7511372  0.8511372  0.9511372  1.05113721
#> 2  02 -0.4079209 -0.3079209 -0.2079209 -0.10792093
#> 3  03 -0.3266974 -0.2266974 -0.1266974 -0.02669743
#> 4  04 -0.4146427 -0.3146427 -0.2146427 -0.11464271
#> 5  05 -0.3792575 -0.2792575 -0.1792575 -0.07925749
#> 6  06  0.0543453  0.1543453  0.2543453  0.35434530
#> 7  07 -1.0076883 -0.9076883 -0.8076883 -0.70768826
#> 8  08 -2.2495010 -2.1495010 -2.0495010 -1.94950097
#> 9  09  0.2477347  0.3477347  0.4477347  0.54773474
#> 10 10  0.2154059  0.3154059  0.4154059  0.51540591

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  2.033914128  2.033914128   2.1339141  2.33391413   2.9339141
#> 2  02  0.006632484  0.006632484   0.1066325  0.30663248   0.9066325
#> 3  03  2.172239121  2.172239121   2.2722391  2.47223912   3.0722391
#> 4  04 -0.203342918 -0.203342918  -0.1033429  0.09665708   0.6966571
#> 5  05  0.230125614  0.230125614   0.3301256  0.53012561   1.1301256
#> 6  06  1.863845336  1.863845336   1.9638453  2.16384534   2.7638453
#> 7  07  0.500890841  0.500890841   0.6008908  0.80089084   1.4008908
#> 8  08 -0.214211445 -0.214211445  -0.1142114  0.08578855   0.6857886
#> 9  09 -0.387532884 -0.387532884  -0.2875329 -0.08753288   0.5124671
#> 10 10  2.185137156  2.185137156   2.2851372  2.48513716   3.0851372