Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.3663695  0.3663695  0.4663695
#> 2  02 -1.5621132 -1.5621132 -1.4621132
#> 3  03  2.1550747  2.1550747  2.2550747
#> 4  04  0.6519000  0.6519000  0.7519000
#> 5  05 -1.6311191 -1.6311191 -1.5311191
#> 6  06 -0.8581162 -0.8581162 -0.7581162
#> 7  07  0.5063127  0.5063127  0.6063127
#> 8  08 -0.3298052 -0.3298052 -0.2298052
#> 9  09  0.5771974  0.5771974  0.6771974
#> 10 10  0.6830966  0.6830966  0.7830966

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1376031 -0.1376031 -0.03760315
#> 2  02  2.3462275  2.3462275  2.44622753
#> 3  03 -1.8205229 -1.8205229 -1.72052290
#> 4  04  0.2707168  0.2707168  0.37071681
#> 5  05  0.2771108  0.2771108  0.37711078
#> 6  06  1.4497394  1.4497394  1.54973942
#> 7  07 -0.4539667 -0.4539667 -0.35396669
#> 8  08 -0.6010040 -0.6010040 -0.50100401
#> 9  09  1.4462525  1.4462525  1.54625252
#> 10 10  1.0668033  1.0668033  1.16680328

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.08792512  0.01207488  0.1120749  0.2120749
#> 2  02  0.70027387  0.80027387  0.9002739  1.0002739
#> 3  03 -0.43557680 -0.33557680 -0.2355768 -0.1355768
#> 4  04 -0.58343595 -0.48343595 -0.3834360 -0.2834360
#> 5  05  0.03324035  0.13324035  0.2332404  0.3332404
#> 6  06  0.58351500  0.68351500  0.7835150  0.8835150
#> 7  07  0.23063542  0.33063542  0.4306354  0.5306354
#> 8  08 -0.92385438 -0.82385438 -0.7238544 -0.6238544
#> 9  09  1.24759184  1.34759184  1.4475918  1.5475918
#> 10 10 -0.06687493  0.03312507  0.1331251  0.2331251

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.5671065  -0.5671065 -0.46710647 -0.26710647   0.3328935
#> 2  02 -0.6316366  -0.6316366 -0.53163657 -0.33163657   0.2683634
#> 3  03 -0.2008690  -0.2008690 -0.10086895  0.09913105   0.6991310
#> 4  04 -1.1164692  -1.1164692 -1.01646922 -0.81646922  -0.2164692
#> 5  05 -0.7314895  -0.7314895 -0.63148955 -0.43148955   0.1685105
#> 6  06 -0.6184885  -0.6184885 -0.51848851 -0.31848851   0.2815115
#> 7  07  0.6798386   0.6798386  0.77983864  0.97983864   1.5798386
#> 8  08 -1.1589095  -1.1589095 -1.05890953 -0.85890953  -0.2589095
#> 9  09 -0.1393438  -0.1393438 -0.03934378  0.16065622   0.7606562
#> 10 10 -1.1299294  -1.1299294 -1.02992945 -0.82992945  -0.2299294