Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  2.3460320  2.3460320  2.44603205
#> 2  02  0.3750334  0.3750334  0.47503339
#> 3  03  0.2317953  0.2317953  0.33179525
#> 4  04 -0.7013055 -0.7013055 -0.60130553
#> 5  05 -0.1367963 -0.1367963 -0.03679633
#> 6  06 -1.7980687 -1.7980687 -1.69806869
#> 7  07  0.3436680  0.3436680  0.44366796
#> 8  08  0.4355161  0.4355161  0.53551614
#> 9  09 -1.1470727 -1.1470727 -1.04707270
#> 10 10  0.2038455  0.2038455  0.30384546

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.97432670 -0.97432670 -0.87432670
#> 2  02 -0.47141883 -0.47141883 -0.37141883
#> 3  03 -0.19373413 -0.19373413 -0.09373413
#> 4  04 -1.19368441 -1.19368441 -1.09368441
#> 5  05 -0.65296912 -0.65296912 -0.55296912
#> 6  06  0.41258413  0.41258413  0.51258413
#> 7  07 -1.82791256 -1.82791256 -1.72791256
#> 8  08 -0.07363437 -0.07363437  0.02636563
#> 9  09 -0.07841951 -0.07841951  0.02158049
#> 10 10  0.70197589  0.70197589  0.80197589

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -0.7402695 -0.64026946 -0.54026946 -0.4402695
#> 2  02  0.8103340  0.91033401  1.01033401  1.1103340
#> 3  03  1.5599116  1.65991161  1.75991161  1.8599116
#> 4  04  0.4199835  0.51998345  0.61998345  0.7199835
#> 5  05  0.7614157  0.86141571  0.96141571  1.0614157
#> 6  06 -1.7847704 -1.68477039 -1.58477039 -1.4847704
#> 7  07 -0.6709613 -0.57096125 -0.47096125 -0.3709613
#> 8  08 -0.1857188 -0.08571881  0.01428119  0.1142812
#> 9  09 -1.1175561 -1.01755606 -0.91755606 -0.8175561
#> 10 10 -1.0385538 -0.93855382 -0.83855382 -0.7385538

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.61635916 -0.61635916  -0.5163592  -0.3163592   0.2836408
#> 2  02  0.40021157  0.40021157   0.5002116   0.7002116   1.3002116
#> 3  03  0.35813660  0.35813660   0.4581366   0.6581366   1.2581366
#> 4  04 -1.35421706 -1.35421706  -1.2542171  -1.0542171  -0.4542171
#> 5  05 -1.43452239 -1.43452239  -1.3345224  -1.1345224  -0.5345224
#> 6  06 -2.17040089 -2.17040089  -2.0704009  -1.8704009  -1.2704009
#> 7  07  1.47243772  1.47243772   1.5724377   1.7724377   2.3724377
#> 8  08  1.47270187  1.47270187   1.5727019   1.7727019   2.3727019
#> 9  09  1.12018533  1.12018533   1.2201853   1.4201853   2.0201853
#> 10 10  0.06598949  0.06598949   0.1659895   0.3659895   0.9659895