Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.2219205 -1.2219205 -1.1219205
#> 2  02  0.1312483  0.1312483  0.2312483
#> 3  03  2.8433008  2.8433008  2.9433008
#> 4  04 -0.5735578 -0.5735578 -0.4735578
#> 5  05 -0.8705224 -0.8705224 -0.7705224
#> 6  06 -0.7583082 -0.7583082 -0.6583082
#> 7  07  0.3844734  0.3844734  0.4844734
#> 8  08  1.1411177  1.1411177  1.2411177
#> 9  09  0.5589694  0.5589694  0.6589694
#> 10 10  0.1462196  0.1462196  0.2462196

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.86598010 -0.86598010 -0.76598010
#> 2  02 -0.15087026 -0.15087026 -0.05087026
#> 3  03 -1.25224760 -1.25224760 -1.15224760
#> 4  04 -0.95612596 -0.95612596 -0.85612596
#> 5  05  1.02471829  1.02471829  1.12471829
#> 6  06 -0.65990825 -0.65990825 -0.55990825
#> 7  07 -0.06472872 -0.06472872  0.03527128
#> 8  08  0.61361908  0.61361908  0.71361908
#> 9  09 -0.39229174 -0.39229174 -0.29229174
#> 10 10  0.28158028  0.28158028  0.38158028

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.5919491 -1.4919491 -1.3919491 -1.2919491
#> 2  02 -0.9550580 -0.8550580 -0.7550580 -0.6550580
#> 3  03 -1.0571144 -0.9571144 -0.8571144 -0.7571144
#> 4  04 -0.7639857 -0.6639857 -0.5639857 -0.4639857
#> 5  05  1.5172580  1.6172580  1.7172580  1.8172580
#> 6  06  0.3670311  0.4670311  0.5670311  0.6670311
#> 7  07 -1.7999246 -1.6999246 -1.5999246 -1.4999246
#> 8  08 -0.6971400 -0.5971400 -0.4971400 -0.3971400
#> 9  09 -1.2717344 -1.1717344 -1.0717344 -0.9717344
#> 10 10  0.1141219  0.2141219  0.3141219  0.4141219

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID             U   Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.5232093189  0.5232093189  0.62320932  0.82320932   1.4232093
#> 2  02  0.9036114975  0.9036114975  1.00361150  1.20361150   1.8036115
#> 3  03  0.3290418877  0.3290418877  0.42904189  0.62904189   1.2290419
#> 4  04 -0.0778543738 -0.0778543738  0.02214563  0.22214563   0.8221456
#> 5  05 -0.0005524291 -0.0005524291  0.09944757  0.29944757   0.8994476
#> 6  06 -0.0836024667 -0.0836024667  0.01639753  0.21639753   0.8163975
#> 7  07 -2.1162874368 -2.1162874368 -2.01628744 -1.81628744  -1.2162874
#> 8  08 -0.2762865408 -0.2762865408 -0.17628654  0.02371346   0.6237135
#> 9  09  0.8206752963  0.8206752963  0.92067530  1.12067530   1.7206753
#> 10 10  1.3510312747  1.3510312747  1.45103127  1.65103127   2.2510313