Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.15799703  0.15799703  0.25799703
#> 2  02 -0.19045501 -0.19045501 -0.09045501
#> 3  03  0.89595170  0.89595170  0.99595170
#> 4  04  0.06225781  0.06225781  0.16225781
#> 5  05 -1.67611866 -1.67611866 -1.57611866
#> 6  06 -0.82052320 -0.82052320 -0.72052320
#> 7  07 -0.12089033 -0.12089033 -0.02089033
#> 8  08  0.46049933  0.46049933  0.56049933
#> 9  09  0.81177462  0.81177462  0.91177462
#> 10 10 -1.21611484 -1.21611484 -1.11611484

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.4277256  1.4277256  1.5277256
#> 2  02  0.8931644  0.8931644  0.9931644
#> 3  03  1.0568572  1.0568572  1.1568572
#> 4  04 -1.4697573 -1.4697573 -1.3697573
#> 5  05 -1.8482380 -1.8482380 -1.7482380
#> 6  06  0.1892741  0.1892741  0.2892741
#> 7  07 -0.4837944 -0.4837944 -0.3837944
#> 8  08  0.6968042  0.6968042  0.7968042
#> 9  09  2.0101705  2.0101705  2.1101705
#> 10 10 -0.6157745 -0.6157745 -0.5157745

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.929766672  1.0297667  1.12976667  1.22976667
#> 2  02  0.005243454  0.1052435  0.20524345  0.30524345
#> 3  03  0.526955728  0.6269557  0.72695573  0.82695573
#> 4  04 -0.240894253 -0.1408943 -0.04089425  0.05910575
#> 5  05 -0.282154451 -0.1821545 -0.08215445  0.01784555
#> 6  06 -2.368656626 -2.2686566 -2.16865663 -2.06865663
#> 7  07  0.348375739  0.4483757  0.54837574  0.64837574
#> 8  08 -1.318378470 -1.2183785 -1.11837847 -1.01837847
#> 9  09 -0.950945175 -0.8509452 -0.75094518 -0.65094518
#> 10 10 -0.240359789 -0.1403598 -0.04035979  0.05964021

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1
#> 1  01 -0.9580142  -0.9580142  -0.8580142 -0.65801418 -0.058014180
#> 2  02 -0.5262753  -0.5262753  -0.4262753 -0.22627533  0.373724669
#> 3  03 -0.4018844  -0.4018844  -0.3018844 -0.10188441  0.498115589
#> 4  04 -1.6125941  -1.6125941  -1.5125941 -1.31259413 -0.712594126
#> 5  05  0.3176391   0.3176391   0.4176391  0.61763915  1.217639147
#> 6  06 -1.3188222  -1.3188222  -1.2188222 -1.01882224 -0.418822236
#> 7  07  0.1058113   0.1058113   0.2058113  0.40581127  1.005811273
#> 8  08 -0.8940327  -0.8940327  -0.7940327 -0.59403272  0.005967279
#> 9  09 -1.5236423  -1.5236423  -1.4236423 -1.22364233 -0.623642334
#> 10 10 -0.3523219  -0.3523219  -0.2523219 -0.05232191  0.547678087