Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.4321679 -1.4321679 -1.3321679
#> 2  02  0.1589307  0.1589307  0.2589307
#> 3  03  0.1272067  0.1272067  0.2272067
#> 4  04 -0.7142121 -0.7142121 -0.6142121
#> 5  05 -0.8193139 -0.8193139 -0.7193139
#> 6  06 -1.0526044 -1.0526044 -0.9526044
#> 7  07  0.9188013  0.9188013  1.0188013
#> 8  08 -0.9192326 -0.9192326 -0.8192326
#> 9  09 -0.8882652 -0.8882652 -0.7882652
#> 10 10  0.3069548  0.3069548  0.4069548

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.11765458  0.11765458  0.2176546
#> 2  02  0.18553595  0.18553595  0.2855360
#> 3  03  0.19957040  0.19957040  0.2995704
#> 4  04  0.44811973  0.44811973  0.5481197
#> 5  05  0.05512825  0.05512825  0.1551283
#> 6  06 -1.58657921 -1.58657921 -1.4865792
#> 7  07 -0.25612890 -0.25612890 -0.1561289
#> 8  08 -1.00066631 -1.00066631 -0.9006663
#> 9  09  0.49015593  0.49015593  0.5901559
#> 10 10  0.36103829  0.36103829  0.4610383

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID         U     Y_Z_1     Y_Z_2     Y_Z_3
#> 1  01 1.9418849 2.0418849 2.1418849 2.2418849
#> 2  02 1.2789662 1.3789662 1.4789662 1.5789662
#> 3  03 1.8948091 1.9948091 2.0948091 2.1948091
#> 4  04 1.2567197 1.3567197 1.4567197 1.5567197
#> 5  05 0.7148541 0.8148541 0.9148541 1.0148541
#> 6  06 0.2584949 0.3584949 0.4584949 0.5584949
#> 7  07 0.5969974 0.6969974 0.7969974 0.8969974
#> 8  08 1.1598115 1.2598115 1.3598115 1.4598115
#> 9  09 0.8399326 0.9399326 1.0399326 1.1399326
#> 10 10 0.2470303 0.3470303 0.4470303 0.5470303

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.420503700  0.420503700  0.52050370  0.72050370   1.3205037
#> 2  02  1.511389082  1.511389082  1.61138908  1.81138908   2.4113891
#> 3  03  0.331169282  0.331169282  0.43116928  0.63116928   1.2311693
#> 4  04  0.279861639  0.279861639  0.37986164  0.57986164   1.1798616
#> 5  05  1.094386528  1.094386528  1.19438653  1.39438653   1.9943865
#> 6  06  0.908359228  0.908359228  1.00835923  1.20835923   1.8083592
#> 7  07 -0.329622236 -0.329622236 -0.22962224 -0.02962224   0.5703778
#> 8  08 -0.005667629 -0.005667629  0.09433237  0.29433237   0.8943324
#> 9  09 -0.562668799 -0.562668799 -0.46266880 -0.26266880   0.3373312
#> 10 10 -0.731810891 -0.731810891 -0.63181089 -0.43181089   0.1681891