Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.02711223 -1.02711223 -0.92711223
#> 2  02 -0.05976544 -0.05976544  0.04023456
#> 3  03 -0.59460509 -0.59460509 -0.49460509
#> 4  04  0.05416196  0.05416196  0.15416196
#> 5  05 -0.97850062 -0.97850062 -0.87850062
#> 6  06  0.09277774  0.09277774  0.19277774
#> 7  07  1.39200234  1.39200234  1.49200234
#> 8  08  1.05243611  1.05243611  1.15243611
#> 9  09 -0.27790847 -0.27790847 -0.17790847
#> 10 10  0.64888140  0.64888140  0.74888140

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.9079010  0.9079010  1.0079010
#> 2  02  1.3143832  1.3143832  1.4143832
#> 3  03 -0.2199553 -0.2199553 -0.1199553
#> 4  04 -0.7964067 -0.7964067 -0.6964067
#> 5  05  0.8767820  0.8767820  0.9767820
#> 6  06  1.0263892  1.0263892  1.1263892
#> 7  07 -1.6809873 -1.6809873 -1.5809873
#> 8  08 -1.3084908 -1.3084908 -1.2084908
#> 9  09 -0.9128615 -0.9128615 -0.8128615
#> 10 10 -0.4015281 -0.4015281 -0.3015281

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.16082824  1.26082824  1.36082824  1.46082824
#> 2  02 -0.27627796 -0.17627796 -0.07627796  0.02372204
#> 3  03 -0.46039121 -0.36039121 -0.26039121 -0.16039121
#> 4  04  0.07237563  0.17237563  0.27237563  0.37237563
#> 5  05  1.12808048  1.22808048  1.32808048  1.42808048
#> 6  06  1.25834849  1.35834849  1.45834849  1.55834849
#> 7  07  0.57339488  0.67339488  0.77339488  0.87339488
#> 8  08 -0.04399126  0.05600874  0.15600874  0.25600874
#> 9  09  0.69462551  0.79462551  0.89462551  0.99462551
#> 10 10  0.26430558  0.36430558  0.46430558  0.56430558

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.31387458  0.31387458  0.413874575   0.6138746  1.21387458
#> 2  02  0.69102275  0.69102275  0.791022746   0.9910227  1.59102275
#> 3  03 -0.10938862 -0.10938862 -0.009388620   0.1906114  0.79061138
#> 4  04 -0.09073383 -0.09073383  0.009266173   0.2092662  0.80926617
#> 5  05  1.53138991  1.53138991  1.631389912   1.8313899  2.43138991
#> 6  06  0.68033869  0.68033869  0.780338689   0.9803387  1.58033869
#> 7  07  0.80864720  0.80864720  0.908647202   1.1086472  1.70864720
#> 8  08 -0.05352967 -0.05352967  0.046470330   0.2464703  0.84647033
#> 9  09 -0.85596324 -0.85596324 -0.755963237  -0.5559632  0.04403676
#> 10 10 -1.38713144 -1.38713144 -1.287131436  -1.0871314 -0.48713144