Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.1293533  1.1293533  1.22935330
#> 2  02 -0.1602625 -0.1602625 -0.06026254
#> 3  03  0.8714735  0.8714735  0.97147350
#> 4  04  0.5043560  0.5043560  0.60435597
#> 5  05 -0.0694628 -0.0694628  0.03053720
#> 6  06  0.8095571  0.8095571  0.90955707
#> 7  07  1.7101452  1.7101452  1.81014515
#> 8  08  0.2035496  0.2035496  0.30354963
#> 9  09  0.5643767  0.5643767  0.66437672
#> 10 10  0.7547622  0.7547622  0.85476224

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.1479420  0.1479420  0.2479420
#> 2  02  0.9732675  0.9732675  1.0732675
#> 3  03  1.3218405  1.3218405  1.4218405
#> 4  04 -0.9883050 -0.9883050 -0.8883050
#> 5  05  0.4416520  0.4416520  0.5416520
#> 6  06  2.0069496  2.0069496  2.1069496
#> 7  07  0.3439750  0.3439750  0.4439750
#> 8  08  0.2163462  0.2163462  0.3163462
#> 9  09  0.2011201  0.2011201  0.3011201
#> 10 10  1.3542855  1.3542855  1.4542855

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -0.62168968 -0.52168968 -0.4216897 -0.32168968
#> 2  02 -0.36319931 -0.26319931 -0.1631993 -0.06319931
#> 3  03 -2.63571137 -2.53571137 -2.4357114 -2.33571137
#> 4  04  0.27261535  0.37261535  0.4726154  0.57261535
#> 5  05  2.74253606  2.84253606  2.9425361  3.04253606
#> 6  06 -0.02440775  0.07559225  0.1755922  0.27559225
#> 7  07 -1.84674847 -1.74674847 -1.6467485 -1.54674847
#> 8  08  0.56099213  0.66099213  0.7609921  0.86099213
#> 9  09  2.00395461  2.10395461  2.2039546  2.30395461
#> 10 10  0.42584513  0.52584513  0.6258451  0.72584513

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.5858800  -1.5858800  -1.4858800  -1.2858800  -0.6858800
#> 2  02  0.1375254   0.1375254   0.2375254   0.4375254   1.0375254
#> 3  03 -1.7970812  -1.7970812  -1.6970812  -1.4970812  -0.8970812
#> 4  04 -1.5430535  -1.5430535  -1.4430535  -1.2430535  -0.6430535
#> 5  05 -1.8097494  -1.8097494  -1.7097494  -1.5097494  -0.9097494
#> 6  06  0.7173995   0.7173995   0.8173995   1.0173995   1.6173995
#> 7  07  0.7544662   0.7544662   0.8544662   1.0544662   1.6544662
#> 8  08  0.3449737   0.3449737   0.4449737   0.6449737   1.2449737
#> 9  09  0.8921166   0.8921166   0.9921166   1.1921166   1.7921166
#> 10 10  0.7047354   0.7047354   0.8047354   1.0047354   1.6047354