Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.0839474  1.0839474  1.18394740
#> 2  02  1.5429316  1.5429316  1.64293159
#> 3  03 -0.1212168 -0.1212168 -0.02121679
#> 4  04  1.0851117  1.0851117  1.18511174
#> 5  05 -0.1925998 -0.1925998 -0.09259975
#> 6  06 -0.4018977 -0.4018977 -0.30189769
#> 7  07  0.5546853  0.5546853  0.65468526
#> 8  08 -0.4577418 -0.4577418 -0.35774183
#> 9  09 -0.1923272 -0.1923272 -0.09232719
#> 10 10  0.3501390  0.3501390  0.45013897

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8056196 -0.8056196 -0.7056196
#> 2  02 -0.9676974 -0.9676974 -0.8676974
#> 3  03  0.5288077  0.5288077  0.6288077
#> 4  04  0.3091955  0.3091955  0.4091955
#> 5  05  0.6370387  0.6370387  0.7370387
#> 6  06  0.2894946  0.2894946  0.3894946
#> 7  07  1.2187114  1.2187114  1.3187114
#> 8  08 -0.5253545 -0.5253545 -0.4253545
#> 9  09  1.0224446  1.0224446  1.1224446
#> 10 10 -1.1888412 -1.1888412 -1.0888412

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.3061067 -1.2061067 -1.1061067 -1.0061067
#> 2  02 -0.9048447 -0.8048447 -0.7048447 -0.6048447
#> 3  03  0.3457811  0.4457811  0.5457811  0.6457811
#> 4  04  0.4285483  0.5285483  0.6285483  0.7285483
#> 5  05 -1.6122863 -1.5122863 -1.4122863 -1.3122863
#> 6  06  0.6275524  0.7275524  0.8275524  0.9275524
#> 7  07 -0.5498914 -0.4498914 -0.3498914 -0.2498914
#> 8  08 -1.1208649 -1.0208649 -0.9208649 -0.8208649
#> 9  09 -1.8636028 -1.7636028 -1.6636028 -1.5636028
#> 10 10  0.7784805  0.8784805  0.9784805  1.0784805

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.2507038   1.2507038   1.3507038  1.55070384   2.1507038
#> 2  02 -1.8920179  -1.8920179  -1.7920179 -1.59201785  -0.9920179
#> 3  03 -1.3849525  -1.3849525  -1.2849525 -1.08495250  -0.4849525
#> 4  04  0.2075774   0.2075774   0.3075774  0.50757742   1.1075774
#> 5  05  2.0999956   2.0999956   2.1999956  2.39999555   2.9999956
#> 6  06 -1.0276563  -1.0276563  -0.9276563 -0.72765626  -0.1276563
#> 7  07  0.1756965   0.1756965   0.2756965  0.47569649   1.0756965
#> 8  08  0.2362424   0.2362424   0.3362424  0.53624236   1.1362424
#> 9  09 -0.3903647  -0.3903647  -0.2903647 -0.09036467   0.5096353
#> 10 10  0.5674430   0.5674430   0.6674430  0.86744295   1.4674430