Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID            U        Y_Z_0       Y_Z_1
#> 1  01  1.645151902  1.645151902  1.74515190
#> 2  02  0.866543175  0.866543175  0.96654317
#> 3  03 -0.008906411 -0.008906411  0.09109359
#> 4  04  0.727519768  0.727519768  0.82751977
#> 5  05  0.222626198  0.222626198  0.32262620
#> 6  06  3.383060426  3.383060426  3.48306043
#> 7  07  0.582189736  0.582189736  0.68218974
#> 8  08 -0.064660796 -0.064660796  0.03533920
#> 9  09 -0.957682094 -0.957682094 -0.85768209
#> 10 10 -1.532679907 -1.532679907 -1.43267991

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.2402561 -0.2402561 -0.1402561
#> 2  02 -1.8305343 -1.8305343 -1.7305343
#> 3  03  1.5074134  1.5074134  1.6074134
#> 4  04 -2.0369143 -2.0369143 -1.9369143
#> 5  05 -0.4405391 -0.4405391 -0.3405391
#> 6  06  0.8120837  0.8120837  0.9120837
#> 7  07 -0.4147744 -0.4147744 -0.3147744
#> 8  08  0.8837189  0.8837189  0.9837189
#> 9  09 -0.8415072 -0.8415072 -0.7415072
#> 10 10 -1.7059134 -1.7059134 -1.6059134

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  1.20789366  1.30789366  1.4078937  1.5078937
#> 2  02  1.33496649  1.43496649  1.5349665  1.6349665
#> 3  03 -1.16585172 -1.06585172 -0.9658517 -0.8658517
#> 4  04  0.12489162  0.22489162  0.3248916  0.4248916
#> 5  05  2.72297363  2.82297363  2.9229736  3.0229736
#> 6  06  0.22960563  0.32960563  0.4296056  0.5296056
#> 7  07 -0.06471745  0.03528255  0.1352826  0.2352826
#> 8  08  2.40020441  2.50020441  2.6002044  2.7002044
#> 9  09  0.39411751  0.49411751  0.5941175  0.6941175
#> 10 10 -1.06507547 -0.96507547 -0.8650755 -0.7650755

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.71176035  0.71176035  0.81176035  1.01176035  1.61176035
#> 2  02  1.01001403  1.01001403  1.11001403  1.31001403  1.91001403
#> 3  03 -0.14955908 -0.14955908 -0.04955908  0.15044092  0.75044092
#> 4  04 -0.09346422 -0.09346422  0.00653578  0.20653578  0.80653578
#> 5  05 -0.31294519 -0.31294519 -0.21294519 -0.01294519  0.58705481
#> 6  06  1.39523149  1.39523149  1.49523149  1.69523149  2.29523149
#> 7  07  0.65970094  0.65970094  0.75970094  0.95970094  1.55970094
#> 8  08 -0.92577337 -0.92577337 -0.82577337 -0.62577337 -0.02577337
#> 9  09 -1.61716192 -1.61716192 -1.51716192 -1.31716192 -0.71716192
#> 10 10  1.05485779  1.05485779  1.15485779  1.35485779  1.95485779