Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.71274305  0.71274305  0.81274305
#> 2  02  0.46032798  0.46032798  0.56032798
#> 3  03  0.07003038  0.07003038  0.17003038
#> 4  04  0.41867030  0.41867030  0.51867030
#> 5  05 -0.14322002 -0.14322002 -0.04322002
#> 6  06 -1.30356717 -1.30356717 -1.20356717
#> 7  07 -1.37232859 -1.37232859 -1.27232859
#> 8  08  0.69044681  0.69044681  0.79044681
#> 9  09 -1.14456576 -1.14456576 -1.04456576
#> 10 10  1.21065718  1.21065718  1.31065718

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.8449039  0.8449039  0.9449039
#> 2  02 -0.2165623 -0.2165623 -0.1165623
#> 3  03 -1.2950048 -1.2950048 -1.1950048
#> 4  04 -0.2727839 -0.2727839 -0.1727839
#> 5  05  0.6274854  0.6274854  0.7274854
#> 6  06  1.0913506  1.0913506  1.1913506
#> 7  07  0.1726827  0.1726827  0.2726827
#> 8  08  0.1315436  0.1315436  0.2315436
#> 9  09  0.5008313  0.5008313  0.6008313
#> 10 10 -0.6812048 -0.6812048 -0.5812048

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.7970984  0.8970984  0.9970984  1.09709835
#> 2  02  1.4908311  1.5908311  1.6908311  1.79083107
#> 3  03  0.7202787  0.8202787  0.9202787  1.02027870
#> 4  04  0.4740854  0.5740854  0.6740854  0.77408539
#> 5  05  1.7102398  1.8102398  1.9102398  2.01023976
#> 6  06  1.0753311  1.1753311  1.2753311  1.37533107
#> 7  07 -0.3429134 -0.2429134 -0.1429134 -0.04291341
#> 8  08  0.3894604  0.4894604  0.5894604  0.68946038
#> 9  09  0.3993474  0.4993474  0.5993474  0.69934745
#> 10 10  1.0854305  1.1854305  1.2854305  1.38543053

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.1501298  -0.1501298 -0.05012983   0.1498702  0.74987017
#> 2  02  0.3280007   0.3280007  0.42800068   0.6280007  1.22800068
#> 3  03  0.7719226   0.7719226  0.87192256   1.0719226  1.67192256
#> 4  04  1.6886935   1.6886935  1.78869347   1.9886935  2.58869347
#> 5  05 -0.4020058  -0.4020058 -0.30200582  -0.1020058  0.49799418
#> 6  06 -0.9432670  -0.9432670 -0.84326696  -0.6432670 -0.04326696
#> 7  07  0.4281928   0.4281928  0.52819278   0.7281928  1.32819278
#> 8  08  0.1167517   0.1167517  0.21675165   0.4167517  1.01675165
#> 9  09 -0.1432468  -0.1432468 -0.04324677   0.1567532  0.75675323
#> 10 10 -1.3633729  -1.3633729 -1.26337292  -1.0633729 -0.46337292