Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  1.08507761  1.08507761  1.1850776
#> 2  02 -1.29911741 -1.29911741 -1.1991174
#> 3  03 -0.97584296 -0.97584296 -0.8758430
#> 4  04  1.60598964  1.60598964  1.7059896
#> 5  05 -0.49928127 -0.49928127 -0.3992813
#> 6  06 -0.44461720 -0.44461720 -0.3446172
#> 7  07 -0.66491549 -0.66491549 -0.5649155
#> 8  08  0.01801928  0.01801928  0.1180193
#> 9  09  0.58169775  0.58169775  0.6816978
#> 10 10 -1.15551182 -1.15551182 -1.0555118

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.4284625 -0.4284625 -0.32846252
#> 2  02  0.9070546  0.9070546  1.00705462
#> 3  03  0.1666969  0.1666969  0.26669686
#> 4  04 -0.1679130 -0.1679130 -0.06791298
#> 5  05 -1.2011916 -1.2011916 -1.10119158
#> 6  06 -1.4760473 -1.4760473 -1.37604731
#> 7  07  0.9385837  0.9385837  1.03858371
#> 8  08 -0.2948827 -0.2948827 -0.19488274
#> 9  09  0.1480567  0.1480567  0.24805670
#> 10 10 -0.3726375 -0.3726375 -0.27263753

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.63419614 -0.53419614 -0.43419614 -0.33419614
#> 2  02 -0.45953701 -0.35953701 -0.25953701 -0.15953701
#> 3  03 -0.46125194 -0.36125194 -0.26125194 -0.16125194
#> 4  04 -0.62459476 -0.52459476 -0.42459476 -0.32459476
#> 5  05  0.04706254  0.14706254  0.24706254  0.34706254
#> 6  06 -0.18070466 -0.08070466  0.01929534  0.11929534
#> 7  07 -0.91976305 -0.81976305 -0.71976305 -0.61976305
#> 8  08 -0.87283825 -0.77283825 -0.67283825 -0.57283825
#> 9  09 -1.38501198 -1.28501198 -1.18501198 -1.08501198
#> 10 10 -0.31277484 -0.21277484 -0.11277484 -0.01277484

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.1709912  -0.1709912 -0.07099118  0.12900882   0.7290088
#> 2  02  1.7313635   1.7313635  1.83136353  2.03136353   2.6313635
#> 3  03  0.1667730   0.1667730  0.26677304  0.46677304   1.0667730
#> 4  04 -0.4139304  -0.4139304 -0.31393044 -0.11393044   0.4860696
#> 5  05  0.7370125   0.7370125  0.83701251  1.03701251   1.6370125
#> 6  06 -0.3771765  -0.3771765 -0.27717654 -0.07717654   0.5228235
#> 7  07  0.6427704   0.6427704  0.74277042  0.94277042   1.5427704
#> 8  08 -1.1468545  -1.1468545 -1.04685450 -0.84685450  -0.2468545
#> 9  09  0.6366826   0.6366826  0.73668259  0.93668259   1.5366826
#> 10 10  1.1499083   1.1499083  1.24990834  1.44990834   2.0499083