Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.5958744  0.5958744  0.69587439
#> 2  02  0.2423129  0.2423129  0.34231294
#> 3  03  0.7512572  0.7512572  0.85125718
#> 4  04 -1.7869906 -1.7869906 -1.68699064
#> 5  05 -0.1290357 -0.1290357 -0.02903566
#> 6  06 -0.1617597 -0.1617597 -0.06175973
#> 7  07  0.6046207  0.6046207  0.70462075
#> 8  08 -0.8449673 -0.8449673 -0.74496734
#> 9  09  0.4426435  0.4426435  0.54264349
#> 10 10 -0.4743829 -0.4743829 -0.37438286

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.1897475 -1.1897475 -1.0897475
#> 2  02  0.5118717  0.5118717  0.6118717
#> 3  03 -0.7754195 -0.7754195 -0.6754195
#> 4  04 -0.4465809 -0.4465809 -0.3465809
#> 5  05 -0.7367338 -0.7367338 -0.6367338
#> 6  06  1.4821268  1.4821268  1.5821268
#> 7  07  1.3474914  1.3474914  1.4474914
#> 8  08 -1.1365648 -1.1365648 -1.0365648
#> 9  09 -1.1964740 -1.1964740 -1.0964740
#> 10 10  0.6879422  0.6879422  0.7879422

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  1.6871961  1.7871961  1.8871961  1.9871961
#> 2  02  0.1766125  0.2766125  0.3766125  0.4766125
#> 3  03  0.9535242  1.0535242  1.1535242  1.2535242
#> 4  04  0.3595538  0.4595538  0.5595538  0.6595538
#> 5  05  0.6130657  0.7130657  0.8130657  0.9130657
#> 6  06 -0.6932229 -0.5932229 -0.4932229 -0.3932229
#> 7  07  0.2669315  0.3669315  0.4669315  0.5669315
#> 8  08  1.3114241  1.4114241  1.5114241  1.6114241
#> 9  09  0.2628222  0.3628222  0.4628222  0.5628222
#> 10 10 -1.2180507 -1.1180507 -1.0180507 -0.9180507

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -2.09633046 -2.09633046 -1.99633046  -1.7963305 -1.19633046
#> 2  02 -0.60639046 -0.60639046 -0.50639046  -0.3063905  0.29360954
#> 3  03 -0.05148256 -0.05148256  0.04851744   0.2485174  0.84851744
#> 4  04  1.14814461  1.14814461  1.24814461   1.4481446  2.04814461
#> 5  05 -1.33206904 -1.33206904 -1.23206904  -1.0320690 -0.43206904
#> 6  06 -0.65486884 -0.65486884 -0.55486884  -0.3548688  0.24513116
#> 7  07 -1.42355701 -1.42355701 -1.32355701  -1.1235570 -0.52355701
#> 8  08  0.30087676  0.30087676  0.40087676   0.6008768  1.20087676
#> 9  09 -0.99865024 -0.99865024 -0.89865024  -0.6986502 -0.09865024
#> 10 10  0.09673535  0.09673535  0.19673535   0.3967354  0.99673535