Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.5807900  0.5807900  0.68078999
#> 2  02  0.2995655  0.2995655  0.39956547
#> 3  03 -0.4437317 -0.4437317 -0.34373175
#> 4  04 -0.2597956 -0.2597956 -0.15979556
#> 5  05  0.5435220  0.5435220  0.64352201
#> 6  06 -0.1771183 -0.1771183 -0.07711833
#> 7  07 -1.0374765 -1.0374765 -0.93747646
#> 8  08  1.1271442  1.1271442  1.22714422
#> 9  09 -0.7206108 -0.7206108 -0.62061080
#> 10 10 -0.6944337 -0.6944337 -0.59443372

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.4537343  0.4537343  0.5537343
#> 2  02  2.2021914  2.2021914  2.3021914
#> 3  03 -1.1017163 -1.1017163 -1.0017163
#> 4  04 -0.1125845 -0.1125845 -0.0125845
#> 5  05 -1.3878037 -1.3878037 -1.2878037
#> 6  06  1.0219958  1.0219958  1.1219958
#> 7  07 -1.1787161 -1.1787161 -1.0787161
#> 8  08  1.4108444  1.4108444  1.5108444
#> 9  09  0.1502742  0.1502742  0.2502742
#> 10 10  1.9698585  1.9698585  2.0698585

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -0.6863483 -0.5863483 -0.4863483 -0.3863483
#> 2  02 -0.9938882 -0.8938882 -0.7938882 -0.6938882
#> 3  03 -0.8881316 -0.7881316 -0.6881316 -0.5881316
#> 4  04 -0.7433457 -0.6433457 -0.5433457 -0.4433457
#> 5  05  0.1295421  0.2295421  0.3295421  0.4295421
#> 6  06  1.4678550  1.5678550  1.6678550  1.7678550
#> 7  07 -1.2577365 -1.1577365 -1.0577365 -0.9577365
#> 8  08  1.3486186  1.4486186  1.5486186  1.6486186
#> 9  09 -0.6211953 -0.5211953 -0.4211953 -0.3211953
#> 10 10  0.0173644  0.1173644  0.2173644  0.3173644

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.8337214   0.8337214   0.9337214  1.13372141   1.7337214
#> 2  02 -0.3781266  -0.3781266  -0.2781266 -0.07812661   0.5218734
#> 3  03  0.6782211   0.6782211   0.7782211  0.97822109   1.5782211
#> 4  04 -0.3557354  -0.3557354  -0.2557354 -0.05573541   0.5442646
#> 5  05 -1.6110771  -1.6110771  -1.5110771 -1.31107706  -0.7110771
#> 6  06  0.2496739   0.2496739   0.3496739  0.54967389   1.1496739
#> 7  07  1.8291436   1.8291436   1.9291436  2.12914360   2.7291436
#> 8  08  0.1426998   0.1426998   0.2426998  0.44269979   1.0426998
#> 9  09 -1.0948817  -1.0948817  -0.9948817 -0.79488173  -0.1948817
#> 10 10 -0.2390030  -0.2390030  -0.1390030  0.06099695   0.6609970