Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.4273700  0.4273700  0.5273700
#> 2  02 -0.4486376 -0.4486376 -0.3486376
#> 3  03 -2.1735579 -2.1735579 -2.0735579
#> 4  04 -0.9133142 -0.9133142 -0.8133142
#> 5  05  0.6088738  0.6088738  0.7088738
#> 6  06 -1.7616915 -1.7616915 -1.6616915
#> 7  07 -0.9233464 -0.9233464 -0.8233464
#> 8  08  0.1252167  0.1252167  0.2252167
#> 9  09  1.3518245  1.3518245  1.4518245
#> 10 10  1.1431867  1.1431867  1.2431867

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.48741038 -0.48741038 -0.3874104
#> 2  02  0.06409264  0.06409264  0.1640926
#> 3  03  0.73808438  0.73808438  0.8380844
#> 4  04  0.74408963  0.74408963  0.8440896
#> 5  05  0.80896050  0.80896050  0.9089605
#> 6  06  0.03731033  0.03731033  0.1373103
#> 7  07 -1.38934879 -1.38934879 -1.2893488
#> 8  08 -0.62467509 -0.62467509 -0.5246751
#> 9  09 -1.25540184 -1.25540184 -1.1554018
#> 10 10 -0.36870739 -0.36870739 -0.2687074

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -0.41899873 -0.31899873 -0.21899873 -0.1189987
#> 2  02  0.09475891  0.19475891  0.29475891  0.3947589
#> 3  03  0.44199570  0.54199570  0.64199570  0.7419957
#> 4  04  1.27373573  1.37373573  1.47373573  1.5737357
#> 5  05  1.76492025  1.86492025  1.96492025  2.0649202
#> 6  06 -0.17343741 -0.07343741  0.02656259  0.1265626
#> 7  07 -0.05942811  0.04057189  0.14057189  0.2405719
#> 8  08  1.16592137  1.26592137  1.36592137  1.4659214
#> 9  09  0.11450166  0.21450166  0.31450166  0.4145017
#> 10 10  0.73156122  0.83156122  0.93156122  1.0315612

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.8815062   1.8815062   1.9815062   2.1815062   2.7815062
#> 2  02  0.4067895   0.4067895   0.5067895   0.7067895   1.3067895
#> 3  03 -0.1856696  -0.1856696  -0.0856696   0.1143304   0.7143304
#> 4  04  0.7138592   0.7138592   0.8138592   1.0138592   1.6138592
#> 5  05 -0.7222778  -0.7222778  -0.6222778  -0.4222778   0.1777222
#> 6  06 -0.6993452  -0.6993452  -0.5993452  -0.3993452   0.2006548
#> 7  07 -0.4518948  -0.4518948  -0.3518948  -0.1518948   0.4481052
#> 8  08  1.5463133   1.5463133   1.6463133   1.8463133   2.4463133
#> 9  09 -1.6272942  -1.6272942  -1.5272942  -1.3272942  -0.7272942
#> 10 10 -0.8904054  -0.8904054  -0.7904054  -0.5904054   0.0095946