Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.2689877  0.2689877  0.3689877
#> 2  02 -1.1994921 -1.1994921 -1.0994921
#> 3  03  1.0019272  1.0019272  1.1019272
#> 4  04  1.1427237  1.1427237  1.2427237
#> 5  05 -1.0862043 -1.0862043 -0.9862043
#> 6  06  2.0089326  2.0089326  2.1089326
#> 7  07 -1.0559148 -1.0559148 -0.9559148
#> 8  08 -0.4110473 -0.4110473 -0.3110473
#> 9  09 -1.0033571 -1.0033571 -0.9033571
#> 10 10  0.3196461  0.3196461  0.4196461

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.8421121  0.8421121  0.94211206
#> 2  02 -0.7341849 -0.7341849 -0.63418486
#> 3  03  0.6329837  0.6329837  0.73298372
#> 4  04 -0.5214850 -0.5214850 -0.42148497
#> 5  05 -1.0275216 -1.0275216 -0.92752159
#> 6  06  0.1055701  0.1055701  0.20557007
#> 7  07  0.3494122  0.3494122  0.44941216
#> 8  08 -1.2375411 -1.2375411 -1.13754112
#> 9  09 -0.5199491 -0.5199491 -0.41994908
#> 10 10 -0.1967640 -0.1967640 -0.09676404

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  0.3611011  0.4611011  0.5611011  0.66110112
#> 2  02 -0.7626581 -0.6626581 -0.5626581 -0.46265814
#> 3  03  1.3119866  1.4119866  1.5119866  1.61198662
#> 4  04 -0.9712284 -0.8712284 -0.7712284 -0.67122839
#> 5  05  0.4875646  0.5875646  0.6875646  0.78756462
#> 6  06 -0.3688093 -0.2688093 -0.1688093 -0.06880931
#> 7  07  0.2749158  0.3749158  0.4749158  0.57491581
#> 8  08  0.6221208  0.7221208  0.8221208  0.92212076
#> 9  09 -0.6655072 -0.5655072 -0.4655072 -0.36550717
#> 10 10  1.2896677  1.3896677  1.4896677  1.58966765

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.5591379  -1.5591379  -1.4591379 -1.25913791 -0.65913791
#> 2  02  2.1155350   2.1155350   2.2155350  2.41553500  3.01553500
#> 3  03 -0.7175293  -0.7175293  -0.6175293 -0.41752927  0.18247073
#> 4  04 -2.5496769  -2.5496769  -2.4496769 -2.24967685 -1.64967685
#> 5  05 -0.2535423  -0.2535423  -0.1535423  0.04645770  0.64645770
#> 6  06  0.2713560   0.2713560   0.3713560  0.57135605  1.17135605
#> 7  07 -0.8673647  -0.8673647  -0.7673647 -0.56736467  0.03263533
#> 8  08 -0.4991754  -0.4991754  -0.3991754 -0.19917537  0.40082463
#> 9  09 -0.3173924  -0.3173924  -0.2173924 -0.01739238  0.58260762
#> 10 10 -2.0210591  -2.0210591  -1.9210591 -1.72105913 -1.12105913