Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.56577281  0.56577281  0.66577281
#> 2  02 -1.03567557 -1.03567557 -0.93567557
#> 3  03  0.06515108  0.06515108  0.16515108
#> 4  04 -0.95466727 -0.95466727 -0.85466727
#> 5  05 -0.38758783 -0.38758783 -0.28758783
#> 6  06 -0.17655742 -0.17655742 -0.07655742
#> 7  07 -1.26283354 -1.26283354 -1.16283354
#> 8  08  0.84065632  0.84065632  0.94065632
#> 9  09  1.57448234  1.57448234  1.67448234
#> 10 10  1.21966617  1.21966617  1.31966617

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  1.3123906  1.3123906  1.41239064
#> 2  02 -0.2029693 -0.2029693 -0.10296930
#> 3  03  0.7986717  0.7986717  0.89867171
#> 4  04  0.4868127  0.4868127  0.58681266
#> 5  05 -1.1664244 -1.1664244 -1.06642439
#> 6  06 -0.1518583 -0.1518583 -0.05185831
#> 7  07 -0.8030072 -0.8030072 -0.70300718
#> 8  08  0.9868345  0.9868345  1.08683445
#> 9  09  1.2163672  1.2163672  1.31636717
#> 10 10 -0.8608143 -0.8608143 -0.76081434

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.03252242  0.1325224  0.2325224  0.3325224
#> 2  02 -0.12286430 -0.0228643  0.0771357  0.1771357
#> 3  03 -1.28251358 -1.1825136 -1.0825136 -0.9825136
#> 4  04 -0.13159660 -0.0315966  0.0684034  0.1684034
#> 5  05  1.26041985  1.3604198  1.4604198  1.5604198
#> 6  06 -0.67240324 -0.5724032 -0.4724032 -0.3724032
#> 7  07  0.90203940  1.0020394  1.1020394  1.2020394
#> 8  08  0.32660801  0.4266080  0.5266080  0.6266080
#> 9  09 -2.02977036 -1.9297704 -1.8297704 -1.7297704
#> 10 10  0.82472586  0.9247259  1.0247259  1.1247259

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.8483912  -0.8483912  -0.7483912  -0.5483912  0.05160878
#> 2  02  1.0461908   1.0461908   1.1461908   1.3461908  1.94619076
#> 3  03 -1.4443716  -1.4443716  -1.3443716  -1.1443716 -0.54437157
#> 4  04  0.2716304   0.2716304   0.3716304   0.5716304  1.17163036
#> 5  05 -0.9702287  -0.9702287  -0.8702287  -0.6702287 -0.07022872
#> 6  06  0.7978364   0.7978364   0.8978364   1.0978364  1.69783637
#> 7  07  1.3590971   1.3590971   1.4590971   1.6590971  2.25909707
#> 8  08  1.3365366   1.3365366   1.4365366   1.6365366  2.23653660
#> 9  09  0.2520593   0.2520593   0.3520593   0.5520593  1.15205929
#> 10 10  0.2806005   0.2806005   0.3806005   0.5806005  1.18060051