Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.6326293  1.6326293  1.7326293
#> 2  02 -1.6886288 -1.6886288 -1.5886288
#> 3  03 -2.0685507 -2.0685507 -1.9685507
#> 4  04 -1.6309478 -1.6309478 -1.5309478
#> 5  05 -0.5145670 -0.5145670 -0.4145670
#> 6  06  0.6199956  0.6199956  0.7199956
#> 7  07 -1.4166294 -1.4166294 -1.3166294
#> 8  08 -1.5179141 -1.5179141 -1.4179141
#> 9  09  1.0484372  1.0484372  1.1484372
#> 10 10 -0.5123476 -0.5123476 -0.4123476

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.3632404 -1.3632404 -1.2632404
#> 2  02  1.3219797  1.3219797  1.4219797
#> 3  03 -1.0577658 -1.0577658 -0.9577658
#> 4  04  0.6057383  0.6057383  0.7057383
#> 5  05 -0.5141175 -0.5141175 -0.4141175
#> 6  06 -0.2547874 -0.2547874 -0.1547874
#> 7  07  1.4881186  1.4881186  1.5881186
#> 8  08 -0.8986803 -0.8986803 -0.7986803
#> 9  09  0.2124857  0.2124857  0.3124857
#> 10 10  0.1055200  0.1055200  0.2055200

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.16039603  0.2603960  0.3603960  0.4603960
#> 2  02  0.03169537  0.1316954  0.2316954  0.3316954
#> 3  03  2.24248054  2.3424805  2.4424805  2.5424805
#> 4  04 -0.78271518 -0.6827152 -0.5827152 -0.4827152
#> 5  05 -0.67691644 -0.5769164 -0.4769164 -0.3769164
#> 6  06 -2.35938226 -2.2593823 -2.1593823 -2.0593823
#> 7  07  0.27918306  0.3791831  0.4791831  0.5791831
#> 8  08  0.48018700  0.5801870  0.6801870  0.7801870
#> 9  09  0.49421633  0.5942163  0.6942163  0.7942163
#> 10 10  0.50155870  0.6015587  0.7015587  0.8015587

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.4206633   0.4206633  0.52066330   0.7206633  1.32066330
#> 2  02 -0.8620947  -0.8620947 -0.76209475  -0.5620947  0.03790525
#> 3  03 -0.5964941  -0.5964941 -0.49649413  -0.2964941  0.30350587
#> 4  04 -1.6622529  -1.6622529 -1.56225289  -1.3622529 -0.76225289
#> 5  05 -0.1960363  -0.1960363 -0.09603631   0.1039637  0.70396369
#> 6  06  0.6586139   0.6586139  0.75861388   0.9586139  1.55861388
#> 7  07  0.6023868   0.6023868  0.70238681   0.9023868  1.50238681
#> 8  08  0.5761273   0.5761273  0.67612730   0.8761273  1.47612730
#> 9  09  0.5926272   0.5926272  0.69262724   0.8926272  1.49262724
#> 10 10  0.1863038   0.1863038  0.28630375   0.4863038  1.08630375