Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.09454085  0.09454085  0.1945409
#> 2  02 -0.43933616 -0.43933616 -0.3393362
#> 3  03 -0.99505907 -0.99505907 -0.8950591
#> 4  04  0.77185390  0.77185390  0.8718539
#> 5  05  2.19036462  2.19036462  2.2903646
#> 6  06  0.78632570  0.78632570  0.8863257
#> 7  07 -0.33812928 -0.33812928 -0.2381293
#> 8  08 -1.45994977 -1.45994977 -1.3599498
#> 9  09  0.29816385  0.29816385  0.3981638
#> 10 10  0.81256590  0.81256590  0.9125659

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.1540137  1.1540137  1.2540137
#> 2  02  1.5959301  1.5959301  1.6959301
#> 3  03 -0.4735761 -0.4735761 -0.3735761
#> 4  04  0.8328732  0.8328732  0.9328732
#> 5  05  0.5165075  0.5165075  0.6165075
#> 6  06 -2.5439879 -2.5439879 -2.4439879
#> 7  07  1.2711661  1.2711661  1.3711661
#> 8  08  0.4391919  0.4391919  0.5391919
#> 9  09  0.7729540  0.7729540  0.8729540
#> 10 10  1.6519095  1.6519095  1.7519095

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01  0.90872873  1.008728729  1.10872873  1.208728729
#> 2  02  0.04236482  0.142364816  0.24236482  0.342364816
#> 3  03 -0.09728170  0.002718302  0.10271830  0.202718302
#> 4  04 -1.23363403 -1.133634030 -1.03363403 -0.933634030
#> 5  05  1.38928582  1.489285818  1.58928582  1.689285818
#> 6  06 -0.16849632 -0.068496320  0.03150368  0.131503680
#> 7  07 -0.25662812 -0.156628119 -0.05662812  0.043371881
#> 8  08  0.92097949  1.020979488  1.12097949  1.220979488
#> 9  09 -0.30889526 -0.208895258 -0.10889526 -0.008895258
#> 10 10  1.73840646  1.838406458  1.93840646  2.038406458

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.21275413  1.21275413  1.312754135   1.5127541   2.1127541
#> 2  02 -0.10488593 -0.10488593 -0.004885933   0.1951141   0.7951141
#> 3  03  0.28956910  0.28956910  0.389569103   0.5895691   1.1895691
#> 4  04  0.15177761  0.15177761  0.251777613   0.4517776   1.0517776
#> 5  05 -1.29638979 -1.29638979 -1.196389788  -0.9963898  -0.3963898
#> 6  06 -0.56440688 -0.56440688 -0.464406884  -0.2644069   0.3355931
#> 7  07 -0.14957277 -0.14957277 -0.049572770   0.1504272   0.7504272
#> 8  08  1.45451018  1.45451018  1.554510180   1.7545102   2.3545102
#> 9  09 -0.07466851 -0.07466851  0.025331492   0.2253315   0.8253315
#> 10 10  0.69528160  0.69528160  0.795281601   0.9952816   1.5952816