Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  2.22683033  2.22683033  2.32683033
#> 2  02  0.31178601  0.31178601  0.41178601
#> 3  03  1.13581114  1.13581114  1.23581114
#> 4  04 -0.59464479 -0.59464479 -0.49464479
#> 5  05 -0.87788032 -0.87788032 -0.77788032
#> 6  06 -0.15742710 -0.15742710 -0.05742710
#> 7  07 -0.01004485 -0.01004485  0.08995515
#> 8  08  0.04034146  0.04034146  0.14034146
#> 9  09 -0.62028019 -0.62028019 -0.52028019
#> 10 10 -0.57369786 -0.57369786 -0.47369786

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.19775912 -0.19775912 -0.09775912
#> 2  02  0.01553649  0.01553649  0.11553649
#> 3  03 -2.77121564 -2.77121564 -2.67121564
#> 4  04  1.03149084  1.03149084  1.13149084
#> 5  05  1.85794748  1.85794748  1.95794748
#> 6  06 -0.08021124 -0.08021124  0.01978876
#> 7  07 -0.62891122 -0.62891122 -0.52891122
#> 8  08 -0.79761391 -0.79761391 -0.69761391
#> 9  09 -0.66640061 -0.66640061 -0.56640061
#> 10 10  1.06324802  1.06324802  1.16324802

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  1.30817380  1.408173800  1.50817380  1.6081738
#> 2  02 -0.17846104 -0.078461037  0.02153896  0.1215390
#> 3  03 -0.09243668  0.007563318  0.10756332  0.2075633
#> 4  04 -0.08073871  0.019261288  0.11926129  0.2192613
#> 5  05 -1.30528979 -1.205289789 -1.10528979 -1.0052898
#> 6  06 -2.51200289 -2.412002894 -2.31200289 -2.2120029
#> 7  07 -1.28608444 -1.186084437 -1.08608444 -0.9860844
#> 8  08  0.07329602  0.173296023  0.27329602  0.3732960
#> 9  09  0.24945592  0.349455916  0.44945592  0.5494559
#> 10 10 -2.18555687 -2.085556869 -1.98555687 -1.8855569

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.6335802   0.6335802   0.7335802  0.93358018   1.5335802
#> 2  02  0.1898279   0.1898279   0.2898279  0.48982790   1.0898279
#> 3  03 -1.8248653  -1.8248653  -1.7248653 -1.52486532  -0.9248653
#> 4  04  1.4345760   1.4345760   1.5345760  1.73457604   2.3345760
#> 5  05 -0.2055409  -0.2055409  -0.1055409  0.09445915   0.6944591
#> 6  06  0.3366498   0.3366498   0.4366498  0.63664976   1.2366498
#> 7  07  0.8074859   0.8074859   0.9074859  1.10748591   1.7074859
#> 8  08  0.6226451   0.6226451   0.7226451  0.92264510   1.5226451
#> 9  09 -1.3756770  -1.3756770  -1.2756770 -1.07567701  -0.4756770
#> 10 10  2.0440364   2.0440364   2.1440364  2.34403640   2.9440364