Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.2252147 -1.2252147 -1.1252147
#> 2  02 -0.7088056 -0.7088056 -0.6088056
#> 3  03 -0.3513069 -0.3513069 -0.2513069
#> 4  04  0.1314793  0.1314793  0.2314793
#> 5  05 -0.4638018 -0.4638018 -0.3638018
#> 6  06 -0.2102265 -0.2102265 -0.1102265
#> 7  07 -0.4557287 -0.4557287 -0.3557287
#> 8  08  0.8172009  0.8172009  0.9172009
#> 9  09 -0.6961677 -0.6961677 -0.5961677
#> 10 10 -0.6805164 -0.6805164 -0.5805164

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.2998446 -0.2998446 -0.1998446
#> 2  02 -0.6139718 -0.6139718 -0.5139718
#> 3  03  0.9626593  0.9626593  1.0626593
#> 4  04 -0.8195105 -0.8195105 -0.7195105
#> 5  05  0.2075542  0.2075542  0.3075542
#> 6  06  0.2964768  0.2964768  0.3964768
#> 7  07  0.9231540  0.9231540  1.0231540
#> 8  08  1.0276894  1.0276894  1.1276894
#> 9  09  0.1862006  0.1862006  0.2862006
#> 10 10 -1.0279752 -1.0279752 -0.9279752

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -2.3392509 -2.2392509 -2.1392509 -2.03925092
#> 2  02 -0.7136355 -0.6136355 -0.5136355 -0.41363548
#> 3  03  1.5413858  1.6413858  1.7413858  1.84138579
#> 4  04 -0.3359257 -0.2359257 -0.1359257 -0.03592574
#> 5  05  0.5729103  0.6729103  0.7729103  0.87291029
#> 6  06 -1.1985489 -1.0985489 -0.9985489 -0.89854893
#> 7  07  0.7799005  0.8799005  0.9799005  1.07990050
#> 8  08  0.4507865  0.5507865  0.6507865  0.75078650
#> 9  09  0.4382443  0.5382443  0.6382443  0.73824429
#> 10 10 -0.8126505 -0.7126505 -0.6126505 -0.51265047

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.806221877  0.806221877  0.90622188  1.10622188  1.70622188
#> 2  02 -1.311701999 -1.311701999 -1.21170200 -1.01170200 -0.41170200
#> 3  03 -0.282484143 -0.282484143 -0.18248414  0.01751586  0.61751586
#> 4  04 -1.072641938 -1.072641938 -0.97264194 -0.77264194 -0.17264194
#> 5  05 -0.953368824 -0.953368824 -0.85336882 -0.65336882 -0.05336882
#> 6  06 -0.003626421 -0.003626421  0.09637358  0.29637358  0.89637358
#> 7  07 -0.053249225 -0.053249225  0.04675078  0.24675078  0.84675078
#> 8  08 -2.191070715 -2.191070715 -2.09107072 -1.89107072 -1.29107072
#> 9  09 -0.788271483 -0.788271483 -0.68827148 -0.48827148  0.11172852
#> 10 10 -1.196520802 -1.196520802 -1.09652080 -0.89652080 -0.29652080