Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.39681776 -0.39681776 -0.2968178
#> 2  02  0.19430043  0.19430043  0.2943004
#> 3  03 -0.58236008 -0.58236008 -0.4823601
#> 4  04  0.05771141  0.05771141  0.1577114
#> 5  05 -1.26349115 -1.26349115 -1.1634912
#> 6  06 -0.27403813 -0.27403813 -0.1740381
#> 7  07  0.25761179  0.25761179  0.3576118
#> 8  08 -0.77112320 -0.77112320 -0.6711232
#> 9  09  0.43818762  0.43818762  0.5381876
#> 10 10  0.29397487  0.29397487  0.3939749

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.05246691  1.05246691  1.15246691
#> 2  02 -0.37042631 -0.37042631 -0.27042631
#> 3  03  0.58165406  0.58165406  0.68165406
#> 4  04 -0.86533358 -0.86533358 -0.76533358
#> 5  05  1.32562092  1.32562092  1.42562092
#> 6  06 -0.33793395 -0.33793395 -0.23793395
#> 7  07 -0.08383792 -0.08383792  0.01616208
#> 8  08  0.20870550  0.20870550  0.30870550
#> 9  09  0.23525442  0.23525442  0.33525442
#> 10 10 -0.03520492 -0.03520492  0.06479508

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.25679447 -0.156794472 -0.05679447  0.04320553
#> 2  02  0.90812301  1.008123009  1.10812301  1.20812301
#> 3  03  0.49752170  0.597521700  0.69752170  0.79752170
#> 4  04 -0.09856922  0.001430779  0.10143078  0.20143078
#> 5  05  1.00233368  1.102333679  1.20233368  1.30233368
#> 6  06  0.02827507  0.128275074  0.22827507  0.32827507
#> 7  07 -1.05114978 -0.951149782 -0.85114978 -0.75114978
#> 8  08  0.90295761  1.002957606  1.10295761  1.20295761
#> 9  09 -2.92052392 -2.820523916 -2.72052392 -2.62052392
#> 10 10  1.65868559  1.758685593  1.85868559  1.95868559

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.84354880  1.84354880  1.943548805   2.1435488   2.7435488
#> 2  02  1.59522655  1.59522655  1.695226552   1.8952266   2.4952266
#> 3  03 -0.76513801 -0.76513801 -0.665138008  -0.4651380   0.1348620
#> 4  04  0.09896201  0.09896201  0.198962005   0.3989620   0.9989620
#> 5  05  3.32944317  3.32944317  3.429443174   3.6294432   4.2294432
#> 6  06  1.90120881  1.90120881  2.001208809   2.2012088   2.8012088
#> 7  07  1.44924569  1.44924569  1.549245692   1.7492457   2.3492457
#> 8  08  0.50566471  0.50566471  0.605664708   0.8056647   1.4056647
#> 9  09 -0.09053362 -0.09053362  0.009466378   0.2094664   0.8094664
#> 10 10 -0.32349850 -0.32349850 -0.223498504  -0.0234985   0.5765015