Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1641850 -0.1641850 -0.06418503
#> 2  02 -0.4844137 -0.4844137 -0.38441369
#> 3  03  0.7469357  0.7469357  0.84693568
#> 4  04 -1.1311128 -1.1311128 -1.03111281
#> 5  05  0.2441270  0.2441270  0.34412702
#> 6  06 -0.6169681 -0.6169681 -0.51696807
#> 7  07 -0.1587349 -0.1587349 -0.05873493
#> 8  08  1.5178453  1.5178453  1.61784528
#> 9  09  0.3595877  0.3595877  0.45958771
#> 10 10  0.3605739  0.3605739  0.46057391

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.41747287 -0.41747287 -0.31747287
#> 2  02 -1.28282274 -1.28282274 -1.18282274
#> 3  03 -0.03406789 -0.03406789  0.06593211
#> 4  04  1.35293315  1.35293315  1.45293315
#> 5  05  0.07664435  0.07664435  0.17664435
#> 6  06 -0.43505234 -0.43505234 -0.33505234
#> 7  07 -0.65042410 -0.65042410 -0.55042410
#> 8  08 -0.17152262 -0.17152262 -0.07152262
#> 9  09 -0.28734906 -0.28734906 -0.18734906
#> 10 10  0.67602809  0.67602809  0.77602809

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.6215472  1.7215472  1.82154724  1.92154724
#> 2  02 -1.3681152 -1.2681152 -1.16811520 -1.06811520
#> 3  03  0.5044790  0.6044790  0.70447900  0.80447900
#> 4  04 -0.8785400 -0.7785400 -0.67853995 -0.57853995
#> 5  05  1.8442290  1.9442290  2.04422897  2.14422897
#> 6  06  0.4870282  0.5870282  0.68702825  0.78702825
#> 7  07 -0.2819687 -0.1819687 -0.08196868  0.01803132
#> 8  08  1.3566672  1.4566672  1.55666718  1.65666718
#> 9  09  0.7661077  0.8661077  0.96610767  1.06610767
#> 10 10  2.1959807  2.2959807  2.39598073  2.49598073

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.43675722 -0.43675722 -0.33675722  -0.1367572   0.4632428
#> 2  02  0.95073932  0.95073932  1.05073932   1.2507393   1.8507393
#> 3  03  0.89846976  0.89846976  0.99846976   1.1984698   1.7984698
#> 4  04 -0.74311457 -0.74311457 -0.64311457  -0.4431146   0.1568854
#> 5  05 -0.13269161 -0.13269161 -0.03269161   0.1673084   0.7673084
#> 6  06 -0.70014677 -0.70014677 -0.60014677  -0.4001468   0.1998532
#> 7  07 -0.01386775 -0.01386775  0.08613225   0.2861323   0.8861323
#> 8  08  0.79487131  0.79487131  0.89487131   1.0948713   1.6948713
#> 9  09 -0.83032590 -0.83032590 -0.73032590  -0.5303259   0.0696741
#> 10 10  0.53269196  0.53269196  0.63269196   0.8326920   1.4326920