Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.3345210  1.3345210  1.4345210
#> 2  02  0.9360663  0.9360663  1.0360663
#> 3  03  0.1350762  0.1350762  0.2350762
#> 4  04  0.7616501  0.7616501  0.8616501
#> 5  05  2.4816311  2.4816311  2.5816311
#> 6  06 -0.9862161 -0.9862161 -0.8862161
#> 7  07 -1.3850817 -1.3850817 -1.2850817
#> 8  08  0.9583863  0.9583863  1.0583863
#> 9  09  0.5719091  0.5719091  0.6719091
#> 10 10  1.6595213  1.6595213  1.7595213

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.49349871  0.49349871  0.59349871
#> 2  02 -0.06482909 -0.06482909  0.03517091
#> 3  03  0.68430617  0.68430617  0.78430617
#> 4  04  0.77714735  0.77714735  0.87714735
#> 5  05 -1.27140940 -1.27140940 -1.17140940
#> 6  06 -0.84174163 -0.84174163 -0.74174163
#> 7  07  0.38793591  0.38793591  0.48793591
#> 8  08  0.77727123  0.77727123  0.87727123
#> 9  09  1.54038738  1.54038738  1.64038738
#> 10 10 -0.07157667 -0.07157667  0.02842333

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.5548220  0.6548220  0.75482196  0.85482196
#> 2  02  0.3367212  0.4367212  0.53672122  0.63672122
#> 3  03 -0.2514822 -0.1514822 -0.05148219  0.04851781
#> 4  04 -0.6338428 -0.5338428 -0.43384284 -0.33384284
#> 5  05 -0.2615919 -0.1615919 -0.06159191  0.03840809
#> 6  06 -0.4251142 -0.3251142 -0.22511417 -0.12511417
#> 7  07 -0.0486231  0.0513769  0.15137690  0.25137690
#> 8  08 -1.6155500 -1.5155500 -1.41555005 -1.31555005
#> 9  09  0.4983112  0.5983112  0.69831122  0.79831122
#> 10 10 -0.5098008 -0.4098008 -0.30980084 -0.20980084

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.3085005  -0.3085005  -0.2085005 -0.008500456   0.5914995
#> 2  02  1.5345503   1.5345503   1.6345503  1.834550336   2.4345503
#> 3  03 -1.7784000  -1.7784000  -1.6784000 -1.478400015  -0.8784000
#> 4  04  0.9494944   0.9494944   1.0494944  1.249494444   1.8494944
#> 5  05  0.1674184   0.1674184   0.2674184  0.467418440   1.0674184
#> 6  06  0.9694362   0.9694362   1.0694362  1.269436246   1.8694362
#> 7  07  0.4702247   0.4702247   0.5702247  0.770224657   1.3702247
#> 8  08 -1.0796311  -1.0796311  -0.9796311 -0.779631105  -0.1796311
#> 9  09  1.4077691   1.4077691   1.5077691  1.707769065   2.3077691
#> 10 10  0.3331489   0.3331489   0.4331489  0.633148901   1.2331489