Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.9968169  0.9968169  1.0968169
#> 2  02 -1.1289552 -1.1289552 -1.0289552
#> 3  03 -0.7476214 -0.7476214 -0.6476214
#> 4  04  0.1493422  0.1493422  0.2493422
#> 5  05 -0.8974839 -0.8974839 -0.7974839
#> 6  06  0.1334117  0.1334117  0.2334117
#> 7  07  0.1257745  0.1257745  0.2257745
#> 8  08  0.7221407  0.7221407  0.8221407
#> 9  09  0.5043257  0.5043257  0.6043257
#> 10 10 -0.4190419 -0.4190419 -0.3190419

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.4759775 -1.4759775 -1.3759775
#> 2  02 -1.9807940 -1.9807940 -1.8807940
#> 3  03  0.2630649  0.2630649  0.3630649
#> 4  04 -0.1731648 -0.1731648 -0.0731648
#> 5  05 -0.8441920 -0.8441920 -0.7441920
#> 6  06 -1.1258745 -1.1258745 -1.0258745
#> 7  07 -0.3816300 -0.3816300 -0.2816300
#> 8  08  0.7743783  0.7743783  0.8743783
#> 9  09 -1.2216509 -1.2216509 -1.1216509
#> 10 10  0.2769674  0.2769674  0.3769674

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.001961792  0.1019618  0.2019618  0.3019618
#> 2  02 -1.094800586 -0.9948006 -0.8948006 -0.7948006
#> 3  03 -0.485396947 -0.3853969 -0.2853969 -0.1853969
#> 4  04 -0.795903848 -0.6959038 -0.5959038 -0.4959038
#> 5  05 -2.422406119 -2.3224061 -2.2224061 -2.1224061
#> 6  06  0.769437046  0.8694370  0.9694370  1.0694370
#> 7  07  0.536481436  0.6364814  0.7364814  0.8364814
#> 8  08  0.439554748  0.5395547  0.6395547  0.7395547
#> 9  09 -1.195784160 -1.0957842 -0.9957842 -0.8957842
#> 10 10 -0.508670919 -0.4086709 -0.3086709 -0.2086709

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1
#> 1  01 -0.0207511  -0.0207511   0.0792489  0.27924890  0.879248901
#> 2  02 -1.7240710  -1.7240710  -1.6240710 -1.42407105 -0.824071046
#> 3  03 -0.2583742  -0.2583742  -0.1583742  0.04162582  0.641625815
#> 4  04 -1.2554735  -1.2554735  -1.1554735 -0.95547350 -0.355473505
#> 5  05 -1.0226678  -1.0226678  -0.9226678 -0.72266779 -0.122667794
#> 6  06 -0.9811081  -0.9811081  -0.8811081 -0.68110811 -0.081108114
#> 7  07  0.4078313   0.4078313   0.5078313  0.70783133  1.307831331
#> 8  08 -1.9028938  -1.9028938  -1.8028938 -1.60289380 -1.002893800
#> 9  09 -0.9066782  -0.9066782  -0.8066782 -0.60667822 -0.006678218
#> 10 10 -2.0226150  -2.0226150  -1.9226150 -1.72261497 -1.122614974