Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.21742763  0.21742763  0.3174276
#> 2  02 -1.00435142 -1.00435142 -0.9043514
#> 3  03 -0.44222566 -0.44222566 -0.3422257
#> 4  04  0.66434278  0.66434278  0.7643428
#> 5  05  0.70841918  0.70841918  0.8084192
#> 6  06 -1.66331180 -1.66331180 -1.5633118
#> 7  07  0.05628254  0.05628254  0.1562825
#> 8  08  1.24610629  1.24610629  1.3461063
#> 9  09  0.82294164  0.82294164  0.9229416
#> 10 10  0.92035624  0.92035624  1.0203562

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.0604195 -1.0604195 -0.9604195
#> 2  02  0.7635379  0.7635379  0.8635379
#> 3  03 -0.5463268 -0.5463268 -0.4463268
#> 4  04 -1.4330512 -1.4330512 -1.3330512
#> 5  05  0.2472703  0.2472703  0.3472703
#> 6  06 -0.1948419 -0.1948419 -0.0948419
#> 7  07  0.3543847  0.3543847  0.4543847
#> 8  08 -0.2732188 -0.2732188 -0.1732188
#> 9  09 -1.3895189 -1.3895189 -1.2895189
#> 10 10  0.4635490  0.4635490  0.5635490

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1        Y_Z_2      Y_Z_3
#> 1  01 -0.5955821 -0.4955821 -0.395582131 -0.2955821
#> 2  02 -1.2911860 -1.1911860 -1.091185952 -0.9911860
#> 3  03 -1.5522214 -1.4522214 -1.352221419 -1.2522214
#> 4  04  0.9165135  1.0165135  1.116513461  1.2165135
#> 5  05  1.1457920  1.2457920  1.345791975  1.4457920
#> 6  06  0.2112268  0.3112268  0.411226791  0.5112268
#> 7  07  0.4772183  0.5772183  0.677218283  0.7772183
#> 8  08  0.9748280  1.0748280  1.174827954  1.2748280
#> 9  09 -0.2080424 -0.1080424 -0.008042404  0.0919576
#> 10 10  0.1514783  0.2514783  0.351478282  0.4514783

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.32601757  0.32601757  0.42601757   0.6260176  1.22601757
#> 2  02  0.16728751  0.16728751  0.26728751   0.4672875  1.06728751
#> 3  03 -0.07778590 -0.07778590  0.02221410   0.2222141  0.82221410
#> 4  04 -0.03164643 -0.03164643  0.06835357   0.2683536  0.86835357
#> 5  05 -0.87806067 -0.87806067 -0.77806067  -0.5780607  0.02193933
#> 6  06 -0.51195266 -0.51195266 -0.41195266  -0.2119527  0.38804734
#> 7  07 -0.70435048 -0.70435048 -0.60435048  -0.4043505  0.19564952
#> 8  08  1.17125888  1.17125888  1.27125888   1.4712589  2.07125888
#> 9  09  0.13597287  0.13597287  0.23597287   0.4359729  1.03597287
#> 10 10  1.74660083  1.74660083  1.84660083   2.0466008  2.64660083