Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID             U         Y_Z_0      Y_Z_1
#> 1  01  1.723393e+00  1.723393e+00  1.8233927
#> 2  02 -8.565979e-01 -8.565979e-01 -0.7565979
#> 3  03 -8.089425e-01 -8.089425e-01 -0.7089425
#> 4  04  3.186287e-01  3.186287e-01  0.4186287
#> 5  05  6.430762e-05  6.430762e-05  0.1000643
#> 6  06  7.453144e-01  7.453144e-01  0.8453144
#> 7  07  7.767634e-01  7.767634e-01  0.8767634
#> 8  08  1.637256e+00  1.637256e+00  1.7372562
#> 9  09  1.909808e-01  1.909808e-01  0.2909808
#> 10 10 -9.642985e-01 -9.642985e-01 -0.8642985

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.78730603 -0.78730603 -0.68730603
#> 2  02 -0.87223292 -0.87223292 -0.77223292
#> 3  03 -0.15281567 -0.15281567 -0.05281567
#> 4  04  0.03013151  0.03013151  0.13013151
#> 5  05 -0.06840230 -0.06840230  0.03159770
#> 6  06  1.53309729  1.53309729  1.63309729
#> 7  07  0.24712297  0.24712297  0.34712297
#> 8  08  0.03180942  0.03180942  0.13180942
#> 9  09  0.02674568  0.02674568  0.12674568
#> 10 10  1.17471354  1.17471354  1.27471354

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  1.6916273  1.79162735  1.89162735  1.9916273
#> 2  02 -0.1232409 -0.02324093  0.07675907  0.1767591
#> 3  03 -1.1225876 -1.02258755 -0.92258755 -0.8225876
#> 4  04  0.4277328  0.52773280  0.62773280  0.7277328
#> 5  05 -0.5997950 -0.49979503 -0.39979503 -0.2997950
#> 6  06 -1.9135684 -1.81356844 -1.71356844 -1.6135684
#> 7  07  2.5269876  2.62698764  2.72698764  2.8269876
#> 8  08  1.0080602  1.10806017  1.20806017  1.3080602
#> 9  09 -0.8719281 -0.77192806 -0.67192806 -0.5719281
#> 10 10 -0.8900331 -0.79003315 -0.69003315 -0.5900331

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.7843570  -0.7843570  -0.6843570  -0.4843570   0.1156430
#> 2  02 -0.4773175  -0.4773175  -0.3773175  -0.1773175   0.4226825
#> 3  03  0.8211899   0.8211899   0.9211899   1.1211899   1.7211899
#> 4  04  0.1086379   0.1086379   0.2086379   0.4086379   1.0086379
#> 5  05  0.1392475   0.1392475   0.2392475   0.4392475   1.0392475
#> 6  06  0.2673954   0.2673954   0.3673954   0.5673954   1.1673954
#> 7  07  1.0425953   1.0425953   1.1425953   1.3425953   1.9425953
#> 8  08  1.0020124   1.0020124   1.1020124   1.3020124   1.9020124
#> 9  09  0.8864158   0.8864158   0.9864158   1.1864158   1.7864158
#> 10 10 -2.0618957  -2.0618957  -1.9618957  -1.7618957  -1.1618957