Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.0257715 -1.0257715 -0.9257715
#> 2  02 -1.8486907 -1.8486907 -1.7486907
#> 3  03 -0.2985646 -0.2985646 -0.1985646
#> 4  04 -0.8473740 -0.8473740 -0.7473740
#> 5  05  1.3134312  1.3134312  1.4134312
#> 6  06 -0.7815953 -0.7815953 -0.6815953
#> 7  07 -0.6502166 -0.6502166 -0.5502166
#> 8  08  0.5044991  0.5044991  0.6044991
#> 9  09  1.0602917  1.0602917  1.1602917
#> 10 10  0.3722642  0.3722642  0.4722642

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -1.13858637 -1.13858637 -1.0385864
#> 2  02  1.34555993  1.34555993  1.4455599
#> 3  03  0.05854821  0.05854821  0.1585482
#> 4  04  0.03945832  0.03945832  0.1394583
#> 5  05 -0.78766676 -0.78766676 -0.6876668
#> 6  06 -1.32317836 -1.32317836 -1.2231784
#> 7  07 -0.23318237 -0.23318237 -0.1331824
#> 8  08 -1.76608536 -1.76608536 -1.6660854
#> 9  09 -0.37502070 -0.37502070 -0.2750207
#> 10 10 -1.21586131 -1.21586131 -1.1158613

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2        Y_Z_3
#> 1  01  0.28662241  0.38662241  0.48662241  0.586622405
#> 2  02  0.80175118  0.90175118  1.00175118  1.101751176
#> 3  03  0.29670027  0.39670027  0.49670027  0.596700268
#> 4  04 -0.50402080 -0.40402080 -0.30402080 -0.204020795
#> 5  05 -0.29721514 -0.19721514 -0.09721514  0.002784857
#> 6  06 -0.31654320 -0.21654320 -0.11654320 -0.016543205
#> 7  07  0.58737017  0.68737017  0.78737017  0.887370167
#> 8  08 -0.03306481  0.06693519  0.16693519  0.266935188
#> 9  09 -2.39278429 -2.29278429 -2.19278429 -2.092784287
#> 10 10  1.18109826  1.28109826  1.38109826  1.481098261

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.105719870  0.105719870   0.2057199  0.40571987   1.0057199
#> 2  02  1.303804869  1.303804869   1.4038049  1.60380487   2.2038049
#> 3  03  0.001411686  0.001411686   0.1014117  0.30141169   0.9014117
#> 4  04 -0.420846707 -0.420846707  -0.3208467 -0.12084671   0.4791533
#> 5  05 -1.198165831 -1.198165831  -1.0981658 -0.89816583  -0.2981658
#> 6  06 -0.355852426 -0.355852426  -0.2558524 -0.05585243   0.5441476
#> 7  07 -0.497968016 -0.497968016  -0.3979680 -0.19796802   0.4020320
#> 8  08 -0.710937797 -0.710937797  -0.6109378 -0.41093780   0.1890622
#> 9  09 -1.657934382 -1.657934382  -1.5579344 -1.35793438  -0.7579344
#> 10 10  0.569464132  0.569464132   0.6694641  0.86946413   1.4694641