Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -1.2530065 -1.2530065 -1.1530065
#> 2  02  1.0029546  1.0029546  1.1029546
#> 3  03  1.3838955  1.3838955  1.4838955
#> 4  04 -0.7036806 -0.7036806 -0.6036806
#> 5  05  0.7216734  0.7216734  0.8216734
#> 6  06  0.7175040  0.7175040  0.8175040
#> 7  07  0.1029086  0.1029086  0.2029086
#> 8  08  1.2130568  1.2130568  1.3130568
#> 9  09 -0.4729125 -0.4729125 -0.3729125
#> 10 10  1.1436523  1.1436523  1.2436523

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  1.24904262  1.24904262  1.34904262
#> 2  02  0.80087030  0.80087030  0.90087030
#> 3  03  0.04011328  0.04011328  0.14011328
#> 4  04  1.89053400  1.89053400  1.99053400
#> 5  05  0.63008733  0.63008733  0.73008733
#> 6  06  0.73050967  0.73050967  0.83050967
#> 7  07  0.20826638  0.20826638  0.30826638
#> 8  08  0.31060064  0.31060064  0.41060064
#> 9  09 -0.65511166 -0.65511166 -0.55511166
#> 10 10 -0.17297026 -0.17297026 -0.07297026

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -0.3669098 -0.2669098 -0.1669098 -0.06690975
#> 2  02  0.5514570  0.6514570  0.7514570  0.85145704
#> 3  03  0.7147905  0.8147905  0.9147905  1.01479053
#> 4  04  0.2699890  0.3699890  0.4699890  0.56998905
#> 5  05 -1.1921926 -1.0921926 -0.9921926 -0.89219259
#> 6  06  2.1681942  2.2681942  2.3681942  2.46819418
#> 7  07 -0.3842045 -0.2842045 -0.1842045 -0.08420446
#> 8  08  0.1496981  0.2496981  0.3496981  0.44969815
#> 9  09 -1.0707199 -0.9707199 -0.8707199 -0.77071991
#> 10 10 -1.8071934 -1.7071934 -1.6071934 -1.50719340

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.16890302 -0.16890302 -0.06890302  0.13109698   0.7310970
#> 2  02  0.89735084  0.89735084  0.99735084  1.19735084   1.7973508
#> 3  03  1.23585329  1.23585329  1.33585329  1.53585329   2.1358533
#> 4  04 -1.26919394 -1.26919394 -1.16919394 -0.96919394  -0.3691939
#> 5  05  1.24219654  1.24219654  1.34219654  1.54219654   2.1421965
#> 6  06 -2.44900211 -2.44900211 -2.34900211 -2.14900211  -1.5490021
#> 7  07 -0.02304978 -0.02304978  0.07695022  0.27695022   0.8769502
#> 8  08 -0.22489458 -0.22489458 -0.12489458  0.07510542   0.6751054
#> 9  09 -1.25555983 -1.25555983 -1.15555983 -0.95555983  -0.3555598
#> 10 10 -1.45040972 -1.45040972 -1.35040972 -1.15040972  -0.5504097