Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8252180 -0.8252180 -0.7252180
#> 2  02  0.2520900  0.2520900  0.3520900
#> 3  03  0.3421744  0.3421744  0.4421744
#> 4  04  0.4059725  0.4059725  0.5059725
#> 5  05  1.0760281  1.0760281  1.1760281
#> 6  06 -0.8100386 -0.8100386 -0.7100386
#> 7  07  0.6694072  0.6694072  0.7694072
#> 8  08 -0.3340215 -0.3340215 -0.2340215
#> 9  09 -0.4336694 -0.4336694 -0.3336694
#> 10 10  1.8078184  1.8078184  1.9078184

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0        Y_Z_1
#> 1  01 -1.0481609 -1.0481609 -0.948160942
#> 2  02  1.1652421  1.1652421  1.265242079
#> 3  03 -0.5435853 -0.5435853 -0.443585262
#> 4  04 -0.1034797 -0.1034797 -0.003479671
#> 5  05 -0.3213374 -0.3213374 -0.221337412
#> 6  06 -1.3824799 -1.3824799 -1.282479871
#> 7  07  1.1041278  1.1041278  1.204127757
#> 8  08  0.6942351  0.6942351  0.794235072
#> 9  09 -0.3687664 -0.3687664 -0.268766384
#> 10 10 -0.4917555 -0.4917555 -0.391755544

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -0.82736418 -0.72736418 -0.62736418 -0.52736418
#> 2  02 -0.76170290 -0.66170290 -0.56170290 -0.46170290
#> 3  03 -0.07177321  0.02822679  0.12822679  0.22822679
#> 4  04  0.64205181  0.74205181  0.84205181  0.94205181
#> 5  05  0.38547089  0.48547089  0.58547089  0.68547089
#> 6  06 -0.27177173 -0.17177173 -0.07177173  0.02822827
#> 7  07 -0.15341356 -0.05341356  0.04658644  0.14658644
#> 8  08 -1.07617930 -0.97617930 -0.87617930 -0.77617930
#> 9  09  0.56876281  0.66876281  0.76876281  0.86876281
#> 10 10  1.67011385  1.77011385  1.87011385  1.97011385

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID             U   Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.9709260460 -1.9709260460  -1.8709260 -1.67092605  -1.0709260
#> 2  02  0.1938945143  0.1938945143   0.2938945  0.49389451   1.0938945
#> 3  03 -0.3951374499 -0.3951374499  -0.2951374 -0.09513745   0.5048626
#> 4  04  1.6381060330  1.6381060330   1.7381060  1.93810603   2.5381060
#> 5  05  0.0009861451  0.0009861451   0.1009861  0.30098615   0.9009861
#> 6  06 -0.2378943289 -0.2378943289  -0.1378943  0.06210567   0.6621057
#> 7  07  0.4945541161  0.4945541161   0.5945541  0.79455412   1.3945541
#> 8  08  0.5872104849  0.5872104849   0.6872105  0.88721048   1.4872105
#> 9  09  0.4621141435  0.4621141435   0.5621141  0.76211414   1.3621141
#> 10 10  0.7772761227  0.7772761227   0.8772761  1.07727612   1.6772761