Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.9258169 -0.9258169 -0.8258169
#> 2  02  0.9647776  0.9647776  1.0647776
#> 3  03  1.2270730  1.2270730  1.3270730
#> 4  04  0.5306409  0.5306409  0.6306409
#> 5  05 -0.4054181 -0.4054181 -0.3054181
#> 6  06  2.3208598  2.3208598  2.4208598
#> 7  07 -0.3579359 -0.3579359 -0.2579359
#> 8  08 -1.3255233 -1.3255233 -1.2255233
#> 9  09  1.2005932  1.2005932  1.3005932
#> 10 10  0.1208961  0.1208961  0.2208961

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID            U        Y_Z_0      Y_Z_1
#> 1  01  0.567369067  0.567369067  0.6673691
#> 2  02  0.002906175  0.002906175  0.1029062
#> 3  03  0.509885556  0.509885556  0.6098856
#> 4  04  0.608957702  0.608957702  0.7089577
#> 5  05  0.003614276  0.003614276  0.1036143
#> 6  06 -0.255153898 -0.255153898 -0.1551539
#> 7  07 -0.599223221 -0.599223221 -0.4992232
#> 8  08  0.179229915  0.179229915  0.2792299
#> 9  09  1.863098607  1.863098607  1.9630986
#> 10 10  0.963293713  0.963293713  1.0632937

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.1242614  0.2242614  0.3242614  0.4242614
#> 2  02  0.1290305  0.2290305  0.3290305  0.4290305
#> 3  03 -2.0796691 -1.9796691 -1.8796691 -1.7796691
#> 4  04  1.1316286  1.2316286  1.3316286  1.4316286
#> 5  05 -0.2012504 -0.1012504 -0.0012504  0.0987496
#> 6  06 -2.6669578 -2.5669578 -2.4669578 -2.3669578
#> 7  07 -1.0199638 -0.9199638 -0.8199638 -0.7199638
#> 8  08  0.2025105  0.3025105  0.4025105  0.5025105
#> 9  09  0.9877522  1.0877522  1.1877522  1.2877522
#> 10 10 -1.2705673 -1.1705673 -1.0705673 -0.9705673

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.83768332 -0.83768332 -0.737683316  -0.5376833  0.06231668
#> 2  02 -0.09264096 -0.09264096  0.007359035   0.2073590  0.80735904
#> 3  03  0.27200872  0.27200872  0.372008718   0.5720087  1.17200872
#> 4  04  0.51282603  0.51282603  0.612826027   0.8128260  1.41282603
#> 5  05  1.21404546  1.21404546  1.314045463   1.5140455  2.11404546
#> 6  06 -1.14905101 -1.14905101 -1.049051010  -0.8490510 -0.24905101
#> 7  07  0.93880708  0.93880708  1.038807082   1.2388071  1.83880708
#> 8  08  1.15787614  1.15787614  1.257876142   1.4578761  2.05787614
#> 9  09  1.12541640  1.12541640  1.225416398   1.4254164  2.02541640
#> 10 10 -0.53029972 -0.53029972 -0.430299716  -0.2302997  0.36970028