Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01 -0.09586482 -0.09586482  0.004135183
#> 2  02  0.27867906  0.27867906  0.378679057
#> 3  03 -0.23497097 -0.23497097 -0.134970965
#> 4  04  0.20378249  0.20378249  0.303782487
#> 5  05  0.09856989  0.09856989  0.198569890
#> 6  06  0.33836181  0.33836181  0.438361813
#> 7  07  1.95754472  1.95754472  2.057544716
#> 8  08  0.37934355  0.37934355  0.479343549
#> 9  09  1.20295597  1.20295597  1.302955968
#> 10 10 -0.45509786 -0.45509786 -0.355097857

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.8377318 -0.8377318 -0.7377318
#> 2  02  1.4709930  1.4709930  1.5709930
#> 3  03  0.7803795  0.7803795  0.8803795
#> 4  04 -0.3264371 -0.3264371 -0.2264371
#> 5  05  0.3865868  0.3865868  0.4865868
#> 6  06  0.9431559  0.9431559  1.0431559
#> 7  07 -0.2076674 -0.2076674 -0.1076674
#> 8  08  0.9980410  0.9980410  1.0980410
#> 9  09  1.5780632  1.5780632  1.6780632
#> 10 10 -0.7938367 -0.7938367 -0.6938367

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01  0.50533124  0.6053312  0.7053312  0.8053312
#> 2  02 -1.21753354 -1.1175335 -1.0175335 -0.9175335
#> 3  03  0.09595186  0.1959519  0.2959519  0.3959519
#> 4  04  1.70441229  1.8044123  1.9044123  2.0044123
#> 5  05  0.34893287  0.4489329  0.5489329  0.6489329
#> 6  06 -0.45538977 -0.3553898 -0.2553898 -0.1553898
#> 7  07  1.12498361  1.2249836  1.3249836  1.4249836
#> 8  08 -0.48022158 -0.3802216 -0.2802216 -0.1802216
#> 9  09  0.20388189  0.3038819  0.4038819  0.5038819
#> 10 10  0.01840832  0.1184083  0.2184083  0.3184083

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.58842815  0.58842815   0.6884282   0.8884282  1.48842815
#> 2  02  0.23311407  0.23311407   0.3331141   0.5331141  1.13311407
#> 3  03 -0.74800207 -0.74800207  -0.6480021  -0.4480021  0.15199793
#> 4  04  0.19481370  0.19481370   0.2948137   0.4948137  1.09481370
#> 5  05  0.88706816  0.88706816   0.9870682   1.1870682  1.78706816
#> 6  06 -0.83533247 -0.83533247  -0.7353325  -0.5353325  0.06466753
#> 7  07  0.07972852  0.07972852   0.1797285   0.3797285  0.97972852
#> 8  08 -0.59602562 -0.59602562  -0.4960256  -0.2960256  0.30397438
#> 9  09 -0.72564479 -0.72564479  -0.6256448  -0.4256448  0.17435521
#> 10 10  0.13787779  0.13787779   0.2378778   0.4378778  1.03787779