Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.83876689 -1.83876689 -1.73876689
#> 2  02 -0.01585474 -0.01585474  0.08414526
#> 3  03  0.22928306  0.22928306  0.32928306
#> 4  04  2.71690395  2.71690395  2.81690395
#> 5  05 -0.41788922 -0.41788922 -0.31788922
#> 6  06  0.96443870  0.96443870  1.06443870
#> 7  07 -0.29282840 -0.29282840 -0.19282840
#> 8  08 -0.31275311 -0.31275311 -0.21275311
#> 9  09  0.60101856  0.60101856  0.70101856
#> 10 10  1.16807676  1.16807676  1.26807676

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01 -0.5338221 -0.5338221 -0.4338221
#> 2  02 -1.2909669 -1.2909669 -1.1909669
#> 3  03  0.3174829  0.3174829  0.4174829
#> 4  04  0.8889049  0.8889049  0.9889049
#> 5  05 -1.1697466 -1.1697466 -1.0697466
#> 6  06 -0.8457114 -0.8457114 -0.7457114
#> 7  07  1.0762009  1.0762009  1.1762009
#> 8  08  0.1209137  0.1209137  0.2209137
#> 9  09 -1.3042231 -1.3042231 -1.2042231
#> 10 10 -0.2415313 -0.2415313 -0.1415313

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U        Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01  0.01349392  0.113493921  0.21349392  0.3134939
#> 2  02 -0.08512140  0.014878596  0.11487860  0.2148786
#> 3  03 -0.10389943 -0.003899425  0.09610057  0.1961006
#> 4  04  1.03837596  1.138375961  1.23837596  1.3383760
#> 5  05 -0.03616312  0.063836877  0.16383688  0.2638369
#> 6  06  0.20669572  0.306695720  0.40669572  0.5066957
#> 7  07 -0.46386682 -0.363866819 -0.26386682 -0.1638668
#> 8  08 -0.71568361 -0.615683609 -0.51568361 -0.4156836
#> 9  09 -0.74170515 -0.641705150 -0.54170515 -0.4417051
#> 10 10 -0.62333353 -0.523333534 -0.42333353 -0.3233335

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  1.81929716  1.81929716  1.91929716   2.1192972   2.7192972
#> 2  02  0.14203009  0.14203009  0.24203009   0.4420301   1.0420301
#> 3  03  0.69232087  0.69232087  0.79232087   0.9923209   1.5923209
#> 4  04  1.39221296  1.39221296  1.49221296   1.6922130   2.2922130
#> 5  05 -0.01596962 -0.01596962  0.08403038   0.2840304   0.8840304
#> 6  06 -1.18820173 -1.18820173 -1.08820173  -0.8882017  -0.2882017
#> 7  07  0.43052057  0.43052057  0.53052057   0.7305206   1.3305206
#> 8  08 -1.35170194 -1.35170194 -1.25170194  -1.0517019  -0.4517019
#> 9  09 -0.53657456 -0.53657456 -0.43657456  -0.2365746   0.3634254
#> 10 10  0.21396971  0.21396971  0.31396971   0.5139697   1.1139697