Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.5295705  1.5295705  1.6295705
#> 2  02 -0.6674304 -0.6674304 -0.5674304
#> 3  03 -3.4303885 -3.4303885 -3.3303885
#> 4  04 -1.7520339 -1.7520339 -1.6520339
#> 5  05  1.9051872  1.9051872  2.0051872
#> 6  06  0.3394899  0.3394899  0.4394899
#> 7  07 -0.8686371 -0.8686371 -0.7686371
#> 8  08  0.7810198  0.7810198  0.8810198
#> 9  09 -0.8527117 -0.8527117 -0.7527117
#> 10 10  0.2369151  0.2369151  0.3369151

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01  0.26152892  0.26152892  0.3615289
#> 2  02 -0.65535778 -0.65535778 -0.5553578
#> 3  03  0.09602369  0.09602369  0.1960237
#> 4  04  0.10225725  0.10225725  0.2022573
#> 5  05 -0.28785329 -0.28785329 -0.1878533
#> 6  06  1.03474764  1.03474764  1.1347476
#> 7  07 -1.63663487 -1.63663487 -1.5366349
#> 8  08  0.24612586  0.24612586  0.3461259
#> 9  09  1.37811555  1.37811555  1.4781155
#> 10 10  1.85602361  1.85602361  1.9560236

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01  1.81218485  1.91218485  2.0121848  2.11218485
#> 2  02 -0.21656250 -0.11656250 -0.0165625  0.08343750
#> 3  03 -0.35436283 -0.25436283 -0.1543628 -0.05436283
#> 4  04 -1.34983910 -1.24983910 -1.1498391 -1.04983910
#> 5  05 -0.05085748  0.04914252  0.1491425  0.24914252
#> 6  06  0.72369039  0.82369039  0.9236904  1.02369039
#> 7  07  2.20726112  2.30726112  2.4072611  2.50726112
#> 8  08  0.75776255  0.85776255  0.9577626  1.05776255
#> 9  09  0.55826898  0.65826898  0.7582690  0.85826898
#> 10 10 -0.03285287  0.06714713  0.1671471  0.26714713

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.8639652  -0.8639652  -0.7639652  -0.5639652  0.03603485
#> 2  02  1.0979313   1.0979313   1.1979313   1.3979313  1.99793127
#> 3  03  1.8382756   1.8382756   1.9382756   2.1382756  2.73827565
#> 4  04  0.7998726   0.7998726   0.8998726   1.0998726  1.69987257
#> 5  05  1.7967166   1.7967166   1.8967166   2.0967166  2.69671659
#> 6  06 -0.6602356  -0.6602356  -0.5602356  -0.3602356  0.23976441
#> 7  07 -1.1657798  -1.1657798  -1.0657798  -0.8657798 -0.26577978
#> 8  08 -1.5493601  -1.5493601  -1.4493601  -1.2493601 -0.64936012
#> 9  09 -1.3464643  -1.3464643  -1.2464643  -1.0464643 -0.44646429
#> 10 10  0.3141900   0.3141900   0.4141900   0.6141900  1.21419005