Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01 -0.70254451 -0.70254451 -0.602544512
#> 2  02  1.87858672  1.87858672  1.978586716
#> 3  03  0.30618825  0.30618825  0.406188252
#> 4  04 -0.10679738 -0.10679738 -0.006797384
#> 5  05 -0.20278323 -0.20278323 -0.102783228
#> 6  06  0.75416970  0.75416970  0.854169704
#> 7  07  0.51905880  0.51905880  0.619058802
#> 8  08 -1.62392356 -1.62392356 -1.523923563
#> 9  09  0.04790147  0.04790147  0.147901469
#> 10 10  1.52114735  1.52114735  1.621147352

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.32513470  0.32513470  0.42513470
#> 2  02  0.34228046  0.34228046  0.44228046
#> 3  03 -0.66909611 -0.66909611 -0.56909611
#> 4  04 -0.17382667 -0.17382667 -0.07382667
#> 5  05 -0.18168719 -0.18168719 -0.08168719
#> 6  06 -0.20834618 -0.20834618 -0.10834618
#> 7  07 -1.66638108 -1.66638108 -1.56638108
#> 8  08 -0.05297546 -0.05297546  0.04702454
#> 9  09  0.70914515  0.70914515  0.80914515
#> 10 10 -0.11429837 -0.11429837 -0.01429837

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U      Y_Z_1      Y_Z_2      Y_Z_3
#> 1  01 -1.10228193 -1.0022819 -0.9022819 -0.8022819
#> 2  02  0.80017179  0.9001718  1.0001718  1.1001718
#> 3  03  1.33548688  1.4354869  1.5354869  1.6354869
#> 4  04  0.02678312  0.1267831  0.2267831  0.3267831
#> 5  05 -1.32996416 -1.2299642 -1.1299642 -1.0299642
#> 6  06  0.02747086  0.1274709  0.2274709  0.3274709
#> 7  07  0.71808778  0.8180878  0.9180878  1.0180878
#> 8  08  0.99707823  1.0970782  1.1970782  1.2970782
#> 9  09  0.68463375  0.7846338  0.8846338  0.9846338
#> 10 10  0.14195719  0.2419572  0.3419572  0.4419572

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -1.7430738  -1.7430738  -1.6430738 -1.44307379  -0.8430738
#> 2  02  1.4616117   1.4616117   1.5616117  1.76161171   2.3616117
#> 3  03  0.7038471   0.7038471   0.8038471  1.00384707   1.6038471
#> 4  04  2.5107156   2.5107156   2.6107156  2.81071562   3.4107156
#> 5  05  0.6876226   0.6876226   0.7876226  0.98762260   1.5876226
#> 6  06 -0.7643304  -0.7643304  -0.6643304 -0.46433044   0.1356696
#> 7  07 -0.2260087  -0.2260087  -0.1260087  0.07399126   0.6739913
#> 8  08 -0.2072467  -0.2072467  -0.1072467  0.09275332   0.6927533
#> 9  09 -0.2293234  -0.2293234  -0.1293234  0.07067658   0.6706766
#> 10 10 -1.4275418  -1.4275418  -1.3275418 -1.12754182  -0.5275418