Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.9712817  1.9712817  2.0712817
#> 2  02 -0.2494287 -0.2494287 -0.1494287
#> 3  03  0.8953015  0.8953015  0.9953015
#> 4  04 -0.5909697 -0.5909697 -0.4909697
#> 5  05 -0.3177558 -0.3177558 -0.2177558
#> 6  06 -1.2339353 -1.2339353 -1.1339353
#> 7  07  1.7054573  1.7054573  1.8054573
#> 8  08 -1.7159687 -1.7159687 -1.6159687
#> 9  09 -0.2389841 -0.2389841 -0.1389841
#> 10 10  0.4426144  0.4426144  0.5426144

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.2844486  0.2844486  0.3844486
#> 2  02 -1.3332950 -1.3332950 -1.2332950
#> 3  03  0.1876546  0.1876546  0.2876546
#> 4  04 -0.4533910 -0.4533910 -0.3533910
#> 5  05 -0.8386864 -0.8386864 -0.7386864
#> 6  06 -0.6373644 -0.6373644 -0.5373644
#> 7  07  0.4498265  0.4498265  0.5498265
#> 8  08 -0.5303771 -0.5303771 -0.4303771
#> 9  09 -0.2971807 -0.2971807 -0.1971807
#> 10 10 -0.4919400 -0.4919400 -0.3919400

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.35847347  0.45847347  0.55847347  0.65847347
#> 2  02 -1.85421940 -1.75421940 -1.65421940 -1.55421940
#> 3  03  0.42564515  0.52564515  0.62564515  0.72564515
#> 4  04 -0.08004148  0.01995852  0.11995852  0.21995852
#> 5  05 -1.25109927 -1.15109927 -1.05109927 -0.95109927
#> 6  06  0.60390922  0.70390922  0.80390922  0.90390922
#> 7  07  0.04511423  0.14511423  0.24511423  0.34511423
#> 8  08 -1.41212313 -1.31212313 -1.21212313 -1.11212313
#> 9  09 -0.23778298 -0.13778298 -0.03778298  0.06221702
#> 10 10 -0.45684288 -0.35684288 -0.25684288 -0.15684288

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.89574665  0.89574665  0.995746650  1.19574665   1.7957467
#> 2  02  0.48283059  0.48283059  0.582830589  0.78283059   1.3828306
#> 3  03  0.96901621  0.96901621  1.069016205  1.26901621   1.8690162
#> 4  04  0.23512476  0.23512476  0.335124763  0.53512476   1.1351248
#> 5  05 -1.77207032 -1.77207032 -1.672070322 -1.47207032  -0.8720703
#> 6  06  0.67446918  0.67446918  0.774469181  0.97446918   1.5744692
#> 7  07 -0.09020541 -0.09020541  0.009794589  0.20979459   0.8097946
#> 8  08  0.72713110  0.72713110  0.827131103  1.02713110   1.6271311
#> 9  09 -0.55985034 -0.55985034 -0.459850335 -0.25985034   0.3401497
#> 10 10 -0.36919205 -0.36919205 -0.269192048 -0.06919205   0.5308080