Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01  0.30507489  0.30507489  0.40507489
#> 2  02  0.68887351  0.68887351  0.78887351
#> 3  03 -0.07136281 -0.07136281  0.02863719
#> 4  04 -0.86605946 -0.86605946 -0.76605946
#> 5  05 -0.49445962 -0.49445962 -0.39445962
#> 6  06 -0.81811432 -0.81811432 -0.71811432
#> 7  07  0.59961520  0.59961520  0.69961520
#> 8  08 -0.68547624 -0.68547624 -0.58547624
#> 9  09  0.44311782  0.44311782  0.54311782
#> 10 10  0.09238352  0.09238352  0.19238352

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -0.02100943 -0.02100943  0.07899057
#> 2  02  0.19434840  0.19434840  0.29434840
#> 3  03 -0.29751294 -0.29751294 -0.19751294
#> 4  04 -1.64456755 -1.64456755 -1.54456755
#> 5  05 -0.63908743 -0.63908743 -0.53908743
#> 6  06 -2.53499003 -2.53499003 -2.43499003
#> 7  07 -0.27906096 -0.27906096 -0.17906096
#> 8  08 -0.47099007 -0.47099007 -0.37099007
#> 9  09 -0.80713155 -0.80713155 -0.70713155
#> 10 10 -0.06612987 -0.06612987  0.03387013

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID           U       Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  1.07089966  1.17089966  1.27089966  1.37089966
#> 2  02 -1.57628918 -1.47628918 -1.37628918 -1.27628918
#> 3  03  0.83432787  0.93432787  1.03432787  1.13432787
#> 4  04 -1.74899476 -1.64899476 -1.54899476 -1.44899476
#> 5  05 -0.46353183 -0.36353183 -0.26353183 -0.16353183
#> 6  06  0.63779433  0.73779433  0.83779433  0.93779433
#> 7  07 -1.17115557 -1.07115557 -0.97115557 -0.87115557
#> 8  08 -0.21272176 -0.11272176 -0.01272176  0.08727824
#> 9  09 -0.06166957  0.03833043  0.13833043  0.23833043
#> 10 10  1.53908344  1.63908344  1.73908344  1.83908344

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.3491645  -0.3491645  -0.2491645 -0.04916448  0.55083552
#> 2  02  1.3976349   1.3976349   1.4976349  1.69763485  2.29763485
#> 3  03  2.2184782   2.2184782   2.3184782  2.51847820  3.11847820
#> 4  04  0.9828525   0.9828525   1.0828525  1.28285247  1.88285247
#> 5  05 -0.3652798  -0.3652798  -0.2652798 -0.06527980  0.53472020
#> 6  06 -1.0552678  -1.0552678  -0.9552678 -0.75526781 -0.15526781
#> 7  07 -0.9125968  -0.9125968  -0.8125968 -0.61259682 -0.01259682
#> 8  08  1.0306872   1.0306872   1.1306872  1.33068720  1.93068720
#> 9  09  1.6567449   1.6567449   1.7567449  1.95674487  2.55674487
#> 10 10 -0.1887540  -0.1887540  -0.0887540  0.11124600  0.71124600