Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0       Y_Z_1
#> 1  01 -1.25379655 -1.25379655 -1.15379655
#> 2  02 -1.09046040 -1.09046040 -0.99046040
#> 3  03  0.05084042  0.05084042  0.15084042
#> 4  04 -0.16620389 -0.16620389 -0.06620389
#> 5  05  1.31115251  1.31115251  1.41115251
#> 6  06  0.90407090  0.90407090  1.00407090
#> 7  07 -0.30493451 -0.30493451 -0.20493451
#> 8  08 -0.36128805 -0.36128805 -0.26128805
#> 9  09  1.63944634  1.63944634  1.73944634
#> 10 10 -0.02214205 -0.02214205  0.07785795

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -0.72495762 -0.72495762 -0.6249576
#> 2  02  0.56793375  0.56793375  0.6679337
#> 3  03  0.83933091  0.83933091  0.9393309
#> 4  04  1.02319108  1.02319108  1.1231911
#> 5  05  0.83603274  0.83603274  0.9360327
#> 6  06 -0.66349118 -0.66349118 -0.5634912
#> 7  07  0.66286822  0.66286822  0.7628682
#> 8  08  1.30016812  1.30016812  1.4001681
#> 9  09 -0.44964913 -0.44964913 -0.3496491
#> 10 10  0.08161318  0.08161318  0.1816132

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1      Y_Z_2       Y_Z_3
#> 1  01 -0.368933536 -0.26893354 -0.1689335 -0.06893354
#> 2  02 -1.017037979 -0.91703798 -0.8170380 -0.71703798
#> 3  03  0.170620393  0.27062039  0.3706204  0.47062039
#> 4  04 -1.027589007 -0.92758901 -0.8275890 -0.72758901
#> 5  05  0.032975363  0.13297536  0.2329754  0.33297536
#> 6  06 -1.078567049 -0.97856705 -0.8785670 -0.77856705
#> 7  07  0.614296734  0.71429673  0.8142967  0.91429673
#> 8  08 -0.008530136  0.09146986  0.1914699  0.29146986
#> 9  09 -0.348435421 -0.24843542 -0.1484354 -0.04843542
#> 10 10  1.089702022  1.18970202  1.2897020  1.38970202

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.34492729  0.34492729  0.444927292  0.644927292  1.24492729
#> 2  02 -0.37949641 -0.37949641 -0.279496413 -0.079496413  0.52050359
#> 3  03  0.90169067  0.90169067  1.001690672  1.201690672  1.80169067
#> 4  04  0.96964963  0.96964963  1.069649633  1.269649633  1.86964963
#> 5  05 -0.80249617 -0.80249617 -0.702496173 -0.502496173  0.09750383
#> 6  06  0.96372616  0.96372616  1.063726165  1.263726165  1.86372616
#> 7  07 -0.11560265 -0.11560265 -0.015602653  0.184397347  0.78439735
#> 8  08 -0.88494264 -0.88494264 -0.784942638 -0.584942638  0.01505736
#> 9  09 -0.09336162 -0.09336162  0.006638378  0.206638378  0.80663838
#> 10 10 -0.30595506 -0.30595506 -0.205955062 -0.005955062  0.59404494