Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  1.7299758  1.7299758  1.8299758
#> 2  02 -2.4566068 -2.4566068 -2.3566068
#> 3  03 -1.1096762 -1.1096762 -1.0096762
#> 4  04  0.3668248  0.3668248  0.4668248
#> 5  05  1.4593559  1.4593559  1.5593559
#> 6  06  0.0759296  0.0759296  0.1759296
#> 7  07 -1.1662596 -1.1662596 -1.0662596
#> 8  08  0.1775330  0.1775330  0.2775330
#> 9  09  1.1997500  1.1997500  1.2997500
#> 10 10  1.2415338  1.2415338  1.3415338

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01  0.2409403  0.2409403  0.34094028
#> 2  02 -0.1576250 -0.1576250 -0.05762497
#> 3  03 -0.9309212 -0.9309212 -0.83092120
#> 4  04  0.2281382  0.2281382  0.32813816
#> 5  05  0.0173464  0.0173464  0.11734640
#> 6  06 -0.9475507 -0.9475507 -0.84755065
#> 7  07  1.0194562  1.0194562  1.11945617
#> 8  08  0.9010788  0.9010788  1.00107875
#> 9  09  0.1736952  0.1736952  0.27369516
#> 10 10  0.4790397  0.4790397  0.57903972

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01  0.4615805  0.5615805  0.66158055  0.76158055
#> 2  02 -0.2190581 -0.1190581 -0.01905808  0.08094192
#> 3  03 -0.5472070 -0.4472070 -0.34720701 -0.24720701
#> 4  04 -0.3324032 -0.2324032 -0.13240317 -0.03240317
#> 5  05 -1.5769515 -1.4769515 -1.37695154 -1.27695154
#> 6  06  0.6665580  0.7665580  0.86655803  0.96655803
#> 7  07 -0.4530488 -0.3530488 -0.25304880 -0.15304880
#> 8  08 -0.3736257 -0.2736257 -0.17362574 -0.07362574
#> 9  09  0.8556525  0.9556525  1.05565245  1.15565245
#> 10 10  0.3754954  0.4754954  0.57549543  0.67549543

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0  Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.2013494   0.2013494   0.3013494  0.501349449  1.10134945
#> 2  02  0.3760046   0.3760046   0.4760046  0.676004617  1.27600462
#> 3  03 -0.2432607  -0.2432607  -0.1432607  0.056739318  0.65673932
#> 4  04  0.7024651   0.7024651   0.8024651  1.002465074  1.60246507
#> 5  05 -0.4142312  -0.4142312  -0.3142312 -0.114231191  0.48576881
#> 6  06  0.4559220   0.4559220   0.5559220  0.755922017  1.35592202
#> 7  07 -1.1570481  -1.1570481  -1.0570481 -0.857048098 -0.25704810
#> 8  08 -0.9186057  -0.9186057  -0.8186057 -0.618605728 -0.01860573
#> 9  09 -0.3075423  -0.3075423  -0.2075423 -0.007542275  0.59245772
#> 10 10  0.1733791   0.1733791   0.2733791  0.473379084  1.07337908