Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0      Y_Z_1
#> 1  01 -1.61064805 -1.61064805 -1.5106481
#> 2  02  0.38383717  0.38383717  0.4838372
#> 3  03  1.40487492  1.40487492  1.5048749
#> 4  04  0.28574870  0.28574870  0.3857487
#> 5  05  0.29972468  0.29972468  0.3997247
#> 6  06  0.17313471  0.17313471  0.2731347
#> 7  07  0.61722639  0.61722639  0.7172264
#> 8  08  1.45731868  1.45731868  1.5573187
#> 9  09  0.05630344  0.05630344  0.1563034
#> 10 10 -0.40177631 -0.40177631 -0.3017763

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0      Y_Z_1
#> 1  01  0.9498299  0.9498299  1.0498299
#> 2  02  0.9200774  0.9200774  1.0200774
#> 3  03 -0.6085168 -0.6085168 -0.5085168
#> 4  04 -0.9138014 -0.9138014 -0.8138014
#> 5  05 -0.2970812 -0.2970812 -0.1970812
#> 6  06 -3.5275939 -3.5275939 -3.4275939
#> 7  07  0.3564422  0.3564422  0.4564422
#> 8  08 -0.5627488 -0.5627488 -0.4627488
#> 9  09 -0.8372150 -0.8372150 -0.7372150
#> 10 10  1.3202758  1.3202758  1.4202758

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID          U      Y_Z_1       Y_Z_2       Y_Z_3
#> 1  01 -1.4968354 -1.3968354 -1.29683536 -1.19683536
#> 2  02  0.4409698  0.5409698  0.64096978  0.74096978
#> 3  03 -0.4591084 -0.3591084 -0.25910842 -0.15910842
#> 4  04 -0.6498511 -0.5498511 -0.44985114 -0.34985114
#> 5  05  0.1257331  0.2257331  0.32573310  0.42573310
#> 6  06  1.1296411  1.2296411  1.32964112  1.42964112
#> 7  07  1.3285618  1.4285618  1.52856181  1.62856181
#> 8  08 -0.2766966 -0.1766966 -0.07669659  0.02330341
#> 9  09  0.8650952  0.9650952  1.06509523  1.16509523
#> 10 10 -0.7332063 -0.6332063 -0.53320630 -0.43320630

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01  0.83589585  0.83589585  0.93589585   1.1358959   1.7358959
#> 2  02  0.14850767  0.14850767  0.24850767   0.4485077   1.0485077
#> 3  03  0.41142275  0.41142275  0.51142275   0.7114227   1.3114227
#> 4  04  1.05767322  1.05767322  1.15767322   1.3576732   1.9576732
#> 5  05 -0.67439518 -0.67439518 -0.57439518  -0.3743952   0.2256048
#> 6  06 -0.50532545 -0.50532545 -0.40532545  -0.2053254   0.3946746
#> 7  07  0.11835798  0.11835798  0.21835798   0.4183580   1.0183580
#> 8  08 -0.03683733 -0.03683733  0.06316267   0.2631627   0.8631627
#> 9  09  1.35768641  1.35768641  1.45768641   1.6576864   2.2576864
#> 10 10  0.29661400  0.29661400  0.39661400   0.5966140   1.1966140