Build potential outcomes variables

Function to draw multiple potential outcomes, one for each condition that an assignment variable can be set to.

Usage

potential_outcomes(x, conditions = list(Z = c(0, 1)), sep = "_")

Arguments

x

Formula describing the potential outcomes with the outcome name on the left hand side and the expression describing the potential outcomes on the right hand side, e.g. Y ~ 0.1 * Z + rnorm(N) (this would draw two potential outcomes columns by default, named Y_Z_0 and Y_Z_1).

conditions

A list of conditions for each assignment variable. Defaults to list(Z = c(0, 1)).

sep

Separator inserted between the outcome name and the assignment variable name used to construct the potential outcome variable names, defaults to "_".

Examples

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)
#>    ID           U       Y_Z_0        Y_Z_1
#> 1  01  0.61349152  0.61349152  0.713491518
#> 2  02 -0.20255430 -0.20255430 -0.102554297
#> 3  03  0.86805023  0.86805023  0.968050231
#> 4  04  1.60039363  1.60039363  1.700393627
#> 5  05 -0.30801763 -0.30801763 -0.208017628
#> 6  06 -0.08415125 -0.08415125  0.015848750
#> 7  07 -0.10583976 -0.10583976 -0.005839759
#> 8  08 -1.72955879 -1.72955879 -1.629558790
#> 9  09 -1.04428050 -1.04428050 -0.944280505
#> 10 10  0.50119085  0.50119085  0.601190852

# equivalently,

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(0, 1)))
)
#>    ID          U      Y_Z_0       Y_Z_1
#> 1  01 -0.1367008 -0.1367008 -0.03670083
#> 2  02 -0.5097682 -0.5097682 -0.40976823
#> 3  03 -1.5809330 -1.5809330 -1.48093297
#> 4  04 -1.8631424 -1.8631424 -1.76314240
#> 5  05 -0.1187139 -0.1187139 -0.01871392
#> 6  06  0.2872976  0.2872976  0.38729764
#> 7  07 -0.3645140 -0.3645140 -0.26451404
#> 8  08 -2.2869990 -2.2869990 -2.18699896
#> 9  09  0.7073549  0.7073549  0.80735495
#> 10 10  0.5845935  0.5845935  0.68459347

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U,
                     conditions = list(Z = c(1, 2, 3)))
)
#>    ID            U       Y_Z_1       Y_Z_2      Y_Z_3
#> 1  01 -0.913172864 -0.81317286 -0.71317286 -0.6131729
#> 2  02  0.890315100  0.99031510  1.09031510  1.1903151
#> 3  03  0.053844559  0.15384456  0.25384456  0.3538446
#> 4  04 -1.759314872 -1.65931487 -1.55931487 -1.4593149
#> 5  05 -0.175529448 -0.07552945  0.02447055  0.1244706
#> 6  06  0.299147440  0.39914744  0.49914744  0.5991474
#> 7  07  0.719437150  0.81943715  0.91943715  1.0194372
#> 8  08 -1.131193683 -1.03119368 -0.93119368 -0.8311937
#> 9  09  0.260906112  0.36090611  0.46090611  0.5609061
#> 10 10 -0.009413096  0.09058690  0.19058690  0.2905869

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1)))
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1
#> 1  01 -0.40464660 -0.40464660  -0.3046466 -0.10464660   0.4953534
#> 2  02  0.58590404  0.58590404   0.6859040  0.88590404   1.4859040
#> 3  03  0.60928306  0.60928306   0.7092831  0.90928306   1.5092831
#> 4  04  1.47676970  1.47676970   1.5767697  1.77676970   2.3767697
#> 5  05  0.07021634  0.07021634   0.1702163  0.37021634   0.9702163
#> 6  06 -2.01422577 -2.01422577  -1.9142258 -1.71422577  -1.1142258
#> 7  07 -0.21801264 -0.21801264  -0.1180126  0.08198736   0.6819874
#> 8  08  1.26111564  1.26111564   1.3611156  1.56111564   2.1611156
#> 9  09 -1.52629058 -1.52629058  -1.4262906 -1.22629058  -0.6262906
#> 10 10  0.68918513  0.68918513   0.7891851  0.98918513   1.5891851