Resample data, including hierarchical data

Resample data, including hierarchical data

resample_data(data, N, ID_labels = NULL)

Arguments

data
A data.frame, usually provided by the user.

N
The number of sample N to return. If multiple levels are defined in ID_labels, N should be a vector of length `length(ID_labels)`

ID_labels
A character vector of the variables that indicate the data hierarchy, from highest to lowest (i.e., from cities to citizens).

Value

A data.frame

Examples

# Bootstrap a dataset without any hierarchy

baseline_survey <- fabricate(N = 5, Y_pre = rnorm(N))
bootsrapped_data <- resample_data(baseline_survey, N = 10)
bootsrapped_data

#>     ID     Y_pre
#> 2    2 0.3764993
#> 4    4 1.2412631
#> 5    5 0.6120909
#> 2.1  2 0.3764993
#> 3    3 1.1387077
#> 2.2  2 0.3764993
#> 2.3  2 0.3764993
#> 2.4  2 0.3764993
#> 1    1 0.1329921
#> 2.5  2 0.3764993

# Bootstrap a hierarchical dataset

my_data <-
fabricate(
  cities = level(N = 2, elevation = runif(n = N, min = 1000, max = 2000)),
  citizens = level(N = 3, income = round(elevation * rnorm(n = N, mean = 5)))
)

my_data_2 <- resample_data(my_data, N = c(3, 5), ID_labels = c("cities", "citizens"))
my_data_2

#>    cities elevation citizens income
#> 1       2  1844.163        6   8593
#> 2       2  1844.163        6   8593
#> 3       2  1844.163        5   7094
#> 4       2  1844.163        5   7094
#> 5       2  1844.163        4  14106
#> 6       1  1526.107        3   4481
#> 7       1  1526.107        3   4481
#> 8       1  1526.107        2   6664
#> 9       1  1526.107        3   4481
#> 10      1  1526.107        1   5954
#> 11      1  1526.107        3   4481
#> 12      1  1526.107        3   4481
#> 13      1  1526.107        3   4481
#> 14      1  1526.107        2   6664
#> 15      1  1526.107        2   6664