Resample data, including hierarchical data

Resample data, including hierarchical data

resample_data(data, N, ID_labels = NULL)

Arguments

data
A data.frame, usually provided by the user.

N
The number of sample N to return. If N is a single scalar and no labels are provided, N will specify the number of unit observations to resample. If N is named, or if the ID_labels argument is specified (in which case, both N and ID_labels should be the same length), then the units resampled will be values of the levels resampled (this is useful for, e.g., cluster resampling). If N is the constant ALL for any level, all units of this level will be transparently passed through to the next level of resampling.

ID_labels
A character vector of the variables that indicate the data hierarchy, from highest to lowest (i.e., from cities to citizens).

Value

A data.frame

Examples

# Bootstrap a dataset without any hierarchy. N specifies a number of observations to return

baseline_survey <- fabricate(N = 5, Y_pre = rnorm(N))
bootsrapped_data <- resample_data(baseline_survey, N = 10)
bootsrapped_data

#>    ID Y_pre
#> 1   5 -0.37
#> 2   1 -1.88
#> 3   1 -1.88
#> 4   3 -2.85
#> 5   1 -1.88
#> 6   3 -2.85
#> 7   1 -1.88
#> 8   3 -2.85
#> 9   3 -2.85
#> 10  1 -1.88

# Resample by a single level of a hierarchical dataset (e.g. resampling clusters of observations)
# N specifies a number of clusters to return

clustered_survey <- fabricate(
  clusters = add_level(N=25),
  cities = add_level(N=round(runif(25, 1, 5)), population=runif(n = N, min=50000, max=1000000))
)

# Specify the name of the cluster variable one of two ways

cluster_resample <- resample_data(clustered_survey, N = 5, ID_labels = "clusters")
cluster_resample

#>    clusters cities population
#> 1        15     45     478992
#> 2        15     46     362256
#> 3        15     47     433650
#> 4        15     48     365767
#> 5        16     49     772547
#> 6        16     50     535382
#> 7        09     28     886871
#> 8        09     29      65558
#> 9        09     30     918230
#> 10       09     31     473302
#> 11       06     19     707426
#> 12       06     20     166994
#> 13       06     21     540668
#> 14       09     28     886871
#> 15       09     29      65558
#> 16       09     30     918230
#> 17       09     31     473302

cluster_resample_2 <- resample_data(clustered_survey, N=c(clusters = 5))
cluster_resample_2

#>    clusters cities population
#> 1        05     15     181036
#> 2        05     16     392431
#> 3        05     17     191511
#> 4        05     18     866554
#> 5        12     37     788833
#> 6        12     38     845408
#> 7        09     28     886871
#> 8        09     29      65558
#> 9        09     30     918230
#> 10       09     31     473302
#> 11       03     06     824049
#> 12       03     07     737258
#> 13       03     08     844693
#> 14       03     09     691182
#> 15       17     51     349820
#> 16       17     52     447777
#> 17       17     53     407002

# Resample a hierarchical dataset on multiple levels

my_data <-
fabricate(
  cities = add_level(N = 2, elevation = runif(n = N, min = 1000, max = 2000)),
  citizens = add_level(N = 3, income = round(elevation * rnorm(n = N, mean = 5)))
)

# Specify the levels you wish to resample one of two ways:
my_data_2 <- resample_data(my_data, N = c(3, 5), ID_labels = c("cities", "citizens"))
my_data_2

#>    cities elevation citizens income
#> 1       2      1170        5   4665
#> 2       2      1170        6   5343
#> 3       2      1170        6   5343
#> 4       2      1170        5   4665
#> 5       2      1170        6   5343
#> 6       1      1775        3   7865
#> 7       1      1775        1  11283
#> 8       1      1775        1  11283
#> 9       1      1775        2   6707
#> 10      1      1775        1  11283
#> 11      2      1170        6   5343
#> 12      2      1170        6   5343
#> 13      2      1170        4   4458
#> 14      2      1170        6   5343
#> 15      2      1170        5   4665

my_data_3 <- resample_data(my_data, N = c(cities=3, citizens=5))
my_data_3

#>    cities elevation citizens income
#> 1       1      1775        2   6707
#> 2       1      1775        2   6707
#> 3       1      1775        3   7865
#> 4       1      1775        3   7865
#> 5       1      1775        3   7865
#> 6       1      1775        2   6707
#> 7       1      1775        3   7865
#> 8       1      1775        1  11283
#> 9       1      1775        2   6707
#> 10      1      1775        2   6707
#> 11      1      1775        2   6707
#> 12      1      1775        2   6707
#> 13      1      1775        1  11283
#> 14      1      1775        1  11283
#> 15      1      1775        2   6707

# Transparently pass through all units at a given level
# This example will resample 2 citizens at each of the cities:

passthrough_resample_data <- resample_data(my_data, N = c(cities=ALL, citizens=2))
passthrough_resample_data

#>   cities elevation citizens income
#> 1      1      1775        2   6707
#> 2      1      1775        1  11283
#> 3      2      1170        5   4665
#> 4      2      1170        5   4665