Resample data, including hierarchical data

Resample data, including hierarchical data

resample_data(data, N, ID_labels = NULL)

Arguments

data
A data.frame, usually provided by the user.

N
The number of sample N to return. If N is a single scalar and no labels are provided, N will specify the number of unit observations to resample. If N is named, or if the ID_labels argument is specified (in which case, both N and ID_labels should be the same length), then the units resampled will be values of the levels resampled (this is useful for, e.g., cluster resampling). If N is the constant ALL for any level, all units of this level will be transparently passed through to the next level of resampling.

ID_labels
A character vector of the variables that indicate the data hierarchy, from highest to lowest (i.e., from cities to citizens).

Value

A data.frame

Examples

# Bootstrap a dataset without any hierarchy. N specifies a number of observations to return

baseline_survey <- fabricate(N = 5, Y_pre = rnorm(N))
bootsrapped_data <- resample_data(baseline_survey, N = 10)
bootsrapped_data

#>     ID     Y_pre
#> 2    2 0.3764993
#> 4    4 1.2412631
#> 5    5 0.6120909
#> 2.1  2 0.3764993
#> 3    3 1.1387077
#> 2.2  2 0.3764993
#> 2.3  2 0.3764993
#> 2.4  2 0.3764993
#> 1    1 0.1329921
#> 2.5  2 0.3764993

# Resample by a single level of a hierarchical dataset (e.g. resampling clusters of observations)
# N specifies a number of clusters to return

clustered_survey <- fabricate(
  clusters = level(N=25),
  cities = level(N=runif(25, 1, 5), population=runif(n = N, min=50000, max=1000000))
)

# Specify the name of the cluster variable one of two ways

cluster_resample <- resample_data(clustered_survey, N = 5, ID_labels = "clusters")
cluster_resample

#>    clusters cities population
#> 21       10     21  372291.19
#> 22       10     22  772723.08
#> 23       10     23   60328.84
#> 24       10     24  109980.47
#> 29       14     29  773192.59
#> 30       14     30  783853.01
#> 15       07     15  329063.25
#> 13       06     13  667030.88
#> 14       06     14  201083.08
#> 4        02     04  835365.36
#> 5        02     05  739475.76
#> 6        02     06  398962.57
#> 7        02     07  547257.00

cluster_resample_2 <- resample_data(clustered_survey, N=c(clusters = 5))
cluster_resample_2

#>    clusters cities population
#> 16       08     16   896720.6
#> 60       25     60   495511.1
#> 61       25     61   892055.5
#> 62       25     62   687021.1
#> 8        03     08   496385.0
#> 38       18     38   239050.7
#> 39       18     39   662802.0
#> 40       18     40   762727.9
#> 41       18     41   307428.9
#> 13       06     13   667030.9
#> 14       06     14   201083.1

# Resample a hierarchical dataset on multiple levels

my_data <-
fabricate(
  cities = level(N = 2, elevation = runif(n = N, min = 1000, max = 2000)),
  citizens = level(N = 3, income = round(elevation * rnorm(n = N, mean = 5)))
)

# Specify the levels you wish to resample one of two ways:
my_data_2 <- resample_data(my_data, N = c(3, 5), ID_labels = c("cities", "citizens"))
my_data_2

#>    cities elevation citizens income
#> 1       2  1927.282        5  10398
#> 2       2  1927.282        4   7181
#> 3       2  1927.282        5  10398
#> 4       2  1927.282        5  10398
#> 5       2  1927.282        4   7181
#> 6       1  1133.349        1   7082
#> 7       1  1133.349        1   7082
#> 8       1  1133.349        3   5500
#> 9       1  1133.349        3   5500
#> 10      1  1133.349        2   4927
#> 11      1  1133.349        1   7082
#> 12      1  1133.349        1   7082
#> 13      1  1133.349        2   4927
#> 14      1  1133.349        2   4927
#> 15      1  1133.349        2   4927

my_data_3 <- resample_data(my_data, N = c(cities=3, citizens=5))
my_data_3

#>    cities elevation citizens income
#> 1       2  1927.282        4   7181
#> 2       2  1927.282        6   5911
#> 3       2  1927.282        5  10398
#> 4       2  1927.282        6   5911
#> 5       2  1927.282        5  10398
#> 6       2  1927.282        5  10398
#> 7       2  1927.282        5  10398
#> 8       2  1927.282        4   7181
#> 9       2  1927.282        6   5911
#> 10      2  1927.282        5  10398
#> 11      1  1133.349        3   5500
#> 12      1  1133.349        2   4927
#> 13      1  1133.349        3   5500
#> 14      1  1133.349        2   4927
#> 15      1  1133.349        3   5500

# Transparently pass through all units at a given level
# This example will resample 2 citizens at each of the cities:

passthrough_resample_data <- resample_data(my_data, N = c(cities=ALL, citizens=2))
passthrough_resample_data

#>   cities elevation citizens income
#> 1      1  1133.349        2   4927
#> 2      1  1133.349        2   4927
#> 3      2  1927.282        5  10398
#> 4      2  1927.282        6   5911