Resample data, including hierarchical data

Resample data, including hierarchical data

resample_data(data, N, ID_labels = NULL)

Arguments

data
A data.frame, usually provided by the user.

N
The number of sample N to return. If N is a single scalar and no labels are provided, N will specify the number of unit observations to resample. If N is named, or if the ID_labels argument is specified (in which case, both N and ID_labels should be the same length), then the units resampled will be values of the levels resampled (this is useful for, e.g., cluster resampling). If N is the constant ALL for any level, all units of this level will be transparently passed through to the next level of resampling.

ID_labels
A character vector of the variables that indicate the data hierarchy, from highest to lowest (i.e., from cities to citizens).

Value

A data.frame

Examples

# Bootstrap a dataset without any hierarchy. N specifies a number of observations to return

baseline_survey <- fabricate(N = 5, Y_pre = rnorm(N))
bootsrapped_data <- resample_data(baseline_survey, N = 10)
bootsrapped_data

#>    ID     Y_pre
#> 1   2 0.3764993
#> 2   4 1.2412631
#> 3   5 0.6120909
#> 4   2 0.3764993
#> 5   3 1.1387077
#> 6   2 0.3764993
#> 7   2 0.3764993
#> 8   2 0.3764993
#> 9   1 0.1329921
#> 10  2 0.3764993

# Resample by a single level of a hierarchical dataset (e.g. resampling clusters of observations)
# N specifies a number of clusters to return

clustered_survey <- fabricate(
  clusters = add_level(N=25),
  cities = add_level(N=round(runif(25, 1, 5)), population=runif(n = N, min=50000, max=1000000))
)

# Specify the name of the cluster variable one of two ways

cluster_resample <- resample_data(clustered_survey, N = 5, ID_labels = "clusters")
cluster_resample

#>    clusters cities population
#> 1        23     62   687021.1
#> 2        23     63   424088.8
#> 3        23     64   573726.1
#> 4        23     65   280368.0
#> 5        06     15   329063.3
#> 6        06     16   896720.6
#> 7        06     17   608674.7
#> 8        07     18   651570.0
#> 9        16     41   307428.9
#> 10       16     42   881840.8
#> 11       16     43   859156.0
#> 12       12     31   431890.9

cluster_resample_2 <- resample_data(clustered_survey, N=c(clusters = 5))
cluster_resample_2

#>    clusters cities population
#> 1        08     19  296534.00
#> 2        03     08  496384.97
#> 3        03     09  143638.07
#> 4        07     18  651569.99
#> 5        17     44   93556.85
#> 6        24     66  243048.39
#> 7        24     67  114924.62
#> 8        24     68  341910.05
#> 9        24     69  993462.19
#> 10       24     70  160493.72

# Resample a hierarchical dataset on multiple levels

my_data <-
fabricate(
  cities = add_level(N = 2, elevation = runif(n = N, min = 1000, max = 2000)),
  citizens = add_level(N = 3, income = round(elevation * rnorm(n = N, mean = 5)))
)

# Specify the levels you wish to resample one of two ways:
my_data_2 <- resample_data(my_data, N = c(3, 5), ID_labels = c("cities", "citizens"))
my_data_2

#>    cities elevation citizens income
#> 1       2  1559.491        5   5589
#> 2       2  1559.491        5   5589
#> 3       2  1559.491        5   5589
#> 4       2  1559.491        6   8745
#> 5       2  1559.491        5   5589
#> 6       1  1026.625        2   4290
#> 7       1  1026.625        1   6216
#> 8       1  1026.625        3   4479
#> 9       1  1026.625        2   4290
#> 10      1  1026.625        3   4479
#> 11      1  1026.625        2   4290
#> 12      1  1026.625        2   4290
#> 13      1  1026.625        2   4290
#> 14      1  1026.625        1   6216
#> 15      1  1026.625        3   4479

my_data_3 <- resample_data(my_data, N = c(cities=3, citizens=5))
my_data_3

#>    cities elevation citizens income
#> 1       2  1559.491        6   8745
#> 2       2  1559.491        5   5589
#> 3       2  1559.491        6   8745
#> 4       2  1559.491        5   5589
#> 5       2  1559.491        5   5589
#> 6       2  1559.491        5   5589
#> 7       2  1559.491        6   8745
#> 8       2  1559.491        5   5589
#> 9       2  1559.491        4   7559
#> 10      2  1559.491        6   8745
#> 11      2  1559.491        6   8745
#> 12      2  1559.491        4   7559
#> 13      2  1559.491        5   5589
#> 14      2  1559.491        6   8745
#> 15      2  1559.491        5   5589

# Transparently pass through all units at a given level
# This example will resample 2 citizens at each of the cities:

passthrough_resample_data <- resample_data(my_data, N = c(cities=ALL, citizens=2))
passthrough_resample_data

#>   cities elevation citizens income
#> 1      1  1026.625        3   4479
#> 2      1  1026.625        3   4479
#> 3      2  1559.491        4   7559
#> 4      2  1559.491        4   7559