Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

reveal_outcomes(x)

Arguments

x
A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)

#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.77895239  0.77895239  0.87895239 1  0.87895239
#> 2  02 -0.42619729 -0.42619729 -0.32619729 1 -0.32619729
#> 3  03 -0.66072292 -0.66072292 -0.56072292 1 -0.56072292
#> 4  04 -1.42540721 -1.42540721 -1.32540721 0 -1.42540721
#> 5  05  0.69742412  0.69742412  0.79742412 1  0.79742412
#> 6  06  0.42597835  0.42597835  0.52597835 1  0.52597835
#> 7  07 -0.07838466 -0.07838466  0.02161534 1  0.02161534
#> 8  08 -0.41904138 -0.41904138 -0.31904138 0 -0.41904138
#> 9  09 -0.32817211 -0.32817211 -0.22817211 0 -0.32817211
#> 10 10 -0.08774028 -0.08774028  0.01225972 0 -0.08774028

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)

#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.4927169   0.4927169  0.59271693   0.7927169   1.3927169  0  0
#> 2  02 -1.2448426  -1.2448426 -1.14484261  -0.9448426  -0.3448426  1  1
#> 3  03 -0.1885761  -0.1885761 -0.08857609   0.1114239   0.7114239  1  0
#> 4  04 -1.0023372  -1.0023372 -0.90233716  -0.7023372  -0.1023372  1  0
#> 5  05  0.5065290   0.5065290  0.60652904   0.8065290   1.4065290  1  0
#> 6  06 -0.5294411  -0.5294411 -0.42944106  -0.2294411   0.3705589  0  1
#> 7  07  0.9512170   0.9512170  1.05121696   1.2512170   1.8512170  1  0
#> 8  08 -1.5344383  -1.5344383 -1.43443826  -1.2344383  -0.6344383  0  0
#> 9  09  2.0428825   2.0428825  2.14288252   2.3428825   2.9428825  1  0
#> 10 10  0.5759681   0.5759681  0.67596815   0.8759681   1.4759681  0  1
#>              Y
#> 1   0.49271693
#> 2  -0.34484261
#> 3  -0.08857609
#> 4  -0.90233716
#> 5   0.60652904
#> 6  -0.22944106
#> 7   1.05121696
#> 8  -1.53443826
#> 9   2.14288252
#> 10  0.87596815