Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z            Y
#> 1  01  0.60406899  0.60406899  0.704068988 0  0.604068988
#> 2  02 -1.18353849 -1.18353849 -1.083538495 1 -1.083538495
#> 3  03 -1.48085250 -1.48085250 -1.380852500 1 -1.380852500
#> 4  04  1.51254349  1.51254349  1.612543488 0  1.512543488
#> 5  05 -0.61651499 -0.61651499 -0.516514994 0 -0.616514994
#> 6  06  1.24619164  1.24619164  1.346191637 0  1.246191637
#> 7  07  0.75497345  0.75497345  0.854973451 1  0.854973451
#> 8  08 -0.79698004 -0.79698004 -0.696980040 1 -0.696980040
#> 9  09 -1.12281736 -1.12281736 -1.022817355 1 -1.022817355
#> 10 10 -0.09555777 -0.09555777  0.004442228 1  0.004442228

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.18730560 -1.18730560  -1.0873056 -0.88730560  -0.2873056  0  0
#> 2  02 -1.64822591 -1.64822591  -1.5482259 -1.34822591  -0.7482259  1  0
#> 3  03 -0.54090889 -0.54090889  -0.4409089 -0.24090889   0.3590911  0  1
#> 4  04  0.38465269  0.38465269   0.4846527  0.68465269   1.2846527  0  0
#> 5  05  0.66529568  0.66529568   0.7652957  0.96529568   1.5652957  0  0
#> 6  06 -0.24210169 -0.24210169  -0.1421017  0.05789831   0.6578983  0  0
#> 7  07 -1.06692988 -1.06692988  -0.9669299 -0.76692988  -0.1669299  0  1
#> 8  08  0.67569928  0.67569928   0.7756993  0.97569928   1.5756993  1  0
#> 9  09  0.27016412  0.27016412   0.3701641  0.57016412   1.1701641  1  0
#> 10 10  0.08795247  0.08795247   0.1879525  0.38795247   0.9879525  1  1
#>             Y
#> 1  -1.1873056
#> 2  -1.5482259
#> 3  -0.2409089
#> 4   0.3846527
#> 5   0.6652957
#> 6  -0.2421017
#> 7  -0.7669299
#> 8   0.7756993
#> 9   0.3701641
#> 10  0.9879525