Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z           Y
#> 1  01  0.75371732  0.75371732  0.853717318 1  0.85371732
#> 2  02  0.35800440  0.35800440  0.458004399 1  0.45800440
#> 3  03 -0.10280835 -0.10280835 -0.002808349 0 -0.10280835
#> 4  04  0.30253805  0.30253805  0.402538053 0  0.30253805
#> 5  05 -0.04498984 -0.04498984  0.055010158 1  0.05501016
#> 6  06 -0.08416026 -0.08416026  0.015839744 1  0.01583974
#> 7  07  0.97697570  0.97697570  1.076975698 0  0.97697570
#> 8  08  1.93313823  1.93313823  2.033138227 1  2.03313823
#> 9  09  0.88071408  0.88071408  0.980714080 0  0.88071408
#> 10 10  0.45456643  0.45456643  0.554566425 0  0.45456643

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.61475731  0.61475731   0.7147573  0.91475731   1.5147573  1  0
#> 2  02  1.02147311  1.02147311   1.1214731  1.32147311   1.9214731  1  0
#> 3  03 -0.65593810 -0.65593810  -0.5559381 -0.35593810   0.2440619  1  1
#> 4  04 -0.28116040 -0.28116040  -0.1811604  0.01883960   0.6188396  1  1
#> 5  05  0.09123858  0.09123858   0.1912386  0.39123858   0.9912386  0  0
#> 6  06  0.99400633  0.99400633   1.0940063  1.29400633   1.8940063  1  1
#> 7  07  0.70660566  0.70660566   0.8066057  1.00660566   1.6066057  0  0
#> 8  08  0.36817982  0.36817982   0.4681798  0.66817982   1.2681798  1  0
#> 9  09  0.34168864  0.34168864   0.4416886  0.64168864   1.2416886  1  0
#> 10 10 -0.38694127 -0.38694127  -0.2869413 -0.08694127   0.5130587  0  1
#>              Y
#> 1   0.71475731
#> 2   1.12147311
#> 3   0.24406190
#> 4   0.61883960
#> 5   0.09123858
#> 6   1.89400633
#> 7   0.70660566
#> 8   0.46817982
#> 9   0.44168864
#> 10 -0.08694127