Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  1.02154333  1.02154333  1.12154333 0  1.02154333
#> 2  02 -0.51988869 -0.51988869 -0.41988869 0 -0.51988869
#> 3  03 -0.06386612 -0.06386612  0.03613388 0 -0.06386612
#> 4  04 -0.16112103 -0.16112103 -0.06112103 1 -0.06112103
#> 5  05  1.99298652  1.99298652  2.09298652 1  2.09298652
#> 6  06 -0.72285642 -0.72285642 -0.62285642 1 -0.62285642
#> 7  07 -1.88772873 -1.88772873 -1.78772873 1 -1.78772873
#> 8  08  1.79732885  1.79732885  1.89732885 1  1.89732885
#> 9  09 -0.39549206 -0.39549206 -0.29549206 0 -0.39549206
#> 10 10  1.53676127  1.53676127  1.63676127 1  1.63676127

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.9757266   0.9757266  1.07572661   1.2757266  1.87572661  0  0
#> 2  02 -0.9234516  -0.9234516 -0.82345162  -0.6234516 -0.02345162  1  0
#> 3  03  0.6468326   0.6468326  0.74683257   0.9468326  1.54683257  0  1
#> 4  04  0.6589504   0.6589504  0.75895039   0.9589504  1.55895039  0  0
#> 5  05 -2.0329843  -2.0329843 -1.93298426  -1.7329843 -1.13298426  1  1
#> 6  06 -1.0915966  -1.0915966 -0.99159657  -0.7915966 -0.19159657  1  0
#> 7  07 -0.5660024  -0.5660024 -0.46600244  -0.2660024  0.33399756  1  0
#> 8  08  0.1913796   0.1913796  0.29137959   0.4913796  1.09137959  0  0
#> 9  09 -0.1353905  -0.1353905 -0.03539053   0.1646095  0.76460947  0  0
#> 10 10 -1.1859105  -1.1859105 -1.08591054  -0.8859105 -0.28591054  0  1
#>             Y
#> 1   0.9757266
#> 2  -0.8234516
#> 3   0.9468326
#> 4   0.6589504
#> 5  -1.1329843
#> 6  -0.9915966
#> 7  -0.4660024
#> 8   0.1913796
#> 9  -0.1353905
#> 10 -0.8859105