Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z           Y
#> 1  01  0.58261593  0.58261593  0.682615929 1  0.68261593
#> 2  02 -1.81639067 -1.81639067 -1.716390668 0 -1.81639067
#> 3  03  1.17032996  1.17032996  1.270329961 1  1.27032996
#> 4  04 -2.02639974 -2.02639974 -1.926399742 0 -2.02639974
#> 5  05  0.68131067  0.68131067  0.781310674 0  0.68131067
#> 6  06 -0.10154946 -0.10154946 -0.001549461 0 -0.10154946
#> 7  07  0.80045065  0.80045065  0.900450652 1  0.90045065
#> 8  08  0.15782518  0.15782518  0.257825183 1  0.25782518
#> 9  09  0.09337039  0.09337039  0.193370389 0  0.09337039
#> 10 10 -1.33591713 -1.33591713 -1.235917133 0 -1.33591713

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.28834896 -1.28834896 -1.18834896  -0.9883490  -0.3883490  1  0
#> 2  02  0.07605542  0.07605542  0.17605542   0.3760554   0.9760554  0  1
#> 3  03  0.76134498  0.76134498  0.86134498   1.0613450   1.6613450  0  0
#> 4  04  0.39241977  0.39241977  0.49241977   0.6924198   1.2924198  0  1
#> 5  05  0.03409661  0.03409661  0.13409661   0.3340966   0.9340966  1  1
#> 6  06 -0.11379432 -0.11379432 -0.01379432   0.1862057   0.7862057  1  0
#> 7  07  2.42077028  2.42077028  2.52077028   2.7207703   3.3207703  0  1
#> 8  08 -1.20322992 -1.20322992 -1.10322992  -0.9032299  -0.3032299  1  0
#> 9  09  0.18579614  0.18579614  0.28579614   0.4857961   1.0857961  1  1
#> 10 10  0.84971035  0.84971035  0.94971035   1.1497103   1.7497103  1  0
#>              Y
#> 1  -1.18834896
#> 2   0.37605542
#> 3   0.76134498
#> 4   0.69241977
#> 5   0.93409661
#> 6  -0.01379432
#> 7   2.72077028
#> 8  -1.10322992
#> 9   1.08579614
#> 10  0.94971035