Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.70339745  0.70339745  0.80339745 1  0.80339745
#> 2  02  0.04748498  0.04748498  0.14748498 0  0.04748498
#> 3  03  0.46475801  0.46475801  0.56475801 1  0.56475801
#> 4  04  0.48900280  0.48900280  0.58900280 1  0.58900280
#> 5  05 -0.51679066 -0.51679066 -0.41679066 1 -0.41679066
#> 6  06  0.52071624  0.52071624  0.62071624 1  0.62071624
#> 7  07  1.64246340  1.64246340  1.74246340 0  1.64246340
#> 8  08  0.82522346  0.82522346  0.92522346 0  0.82522346
#> 9  09  0.76051630  0.76051630  0.86051630 1  0.86051630
#> 10 10 -0.08070798 -0.08070798  0.01929202 0 -0.08070798

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.6819478  -0.6819478  -0.5819478 -0.38194783  0.21805217  0  1
#> 2  02  1.4861104   1.4861104   1.5861104  1.78611038  2.38611038  1  1
#> 3  03 -0.8488155  -0.8488155  -0.7488155 -0.54881549  0.05118451  1  0
#> 4  04 -0.2117721  -0.2117721  -0.1117721  0.08822791  0.68822791  0  0
#> 5  05  0.2977089   0.2977089   0.3977089  0.59770888  1.19770888  0  1
#> 6  06  0.0250050   0.0250050   0.1250050  0.32500500  0.92500500  0  0
#> 7  07 -0.9400382  -0.9400382  -0.8400382 -0.64003823 -0.04003823  0  1
#> 8  08 -0.6997290  -0.6997290  -0.5997290 -0.39972900  0.20027100  0  0
#> 9  09  1.1940149   1.1940149   1.2940149  1.49401492  2.09401492  0  0
#> 10 10 -2.0033390  -2.0033390  -1.9033390 -1.70333895 -1.10333895  0  1
#>             Y
#> 1  -0.3819478
#> 2   2.3861104
#> 3  -0.7488155
#> 4  -0.2117721
#> 5   0.5977089
#> 6   0.0250050
#> 7  -0.6400382
#> 8  -0.6997290
#> 9   1.1940149
#> 10 -1.7033390