Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z          Y
#> 1  01 -0.4840042 -0.4840042 -0.38400418 1 -0.3840042
#> 2  02 -0.5805300 -0.5805300 -0.48052998 0 -0.5805300
#> 3  03 -0.2199526 -0.2199526 -0.11995258 1 -0.1199526
#> 4  04  0.4678609  0.4678609  0.56786087 1  0.5678609
#> 5  05 -0.1749322 -0.1749322 -0.07493215 0 -0.1749322
#> 6  06 -0.8677973 -0.8677973 -0.76779729 1 -0.7677973
#> 7  07  0.7745904  0.7745904  0.87459035 0  0.7745904
#> 8  08  1.6011265  1.6011265  1.70112654 1  1.7011265
#> 9  09 -0.2311126 -0.2311126 -0.13111256 0 -0.2311126
#> 10 10 -0.7088577 -0.7088577 -0.60885767 0 -0.7088577

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.5351015   0.5351015   0.6351015  0.83510151   1.4351015  0  1
#> 2  02 -0.6408046  -0.6408046  -0.5408046 -0.34080464   0.2591954  0  0
#> 3  03  0.2045772   0.2045772   0.3045772  0.50457721   1.1045772  0  1
#> 4  04  0.4331671   0.4331671   0.5331671  0.73316712   1.3331671  0  0
#> 5  05  0.6936700   0.6936700   0.7936700  0.99367000   1.5936700  0  0
#> 6  06  0.2453001   0.2453001   0.3453001  0.54530014   1.1453001  0  0
#> 7  07  1.1523070   1.1523070   1.2523070  1.45230701   2.0523070  1  0
#> 8  08 -0.7982380  -0.7982380  -0.6982380 -0.49823799   0.1017620  0  0
#> 9  09 -0.2572794  -0.2572794  -0.1572794  0.04272057   0.6427206  0  0
#> 10 10 -1.3195252  -1.3195252  -1.2195252 -1.01952520  -0.4195252  0  1
#>             Y
#> 1   0.8351015
#> 2  -0.6408046
#> 3   0.5045772
#> 4   0.4331671
#> 5   0.6936700
#> 6   0.2453001
#> 7   1.2523070
#> 8  -0.7982380
#> 9  -0.2572794
#> 10 -1.0195252