Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.33771023 -0.33771023 -0.23771023 1 -0.23771023
#> 2  02  0.12609319  0.12609319  0.22609319 0  0.12609319
#> 3  03 -0.03959584 -0.03959584  0.06040416 0 -0.03959584
#> 4  04 -0.24054791 -0.24054791 -0.14054791 1 -0.14054791
#> 5  05 -0.36438663 -0.36438663 -0.26438663 1 -0.26438663
#> 6  06  1.15780741  1.15780741  1.25780741 1  1.25780741
#> 7  07  0.15319296  0.15319296  0.25319296 1  0.25319296
#> 8  08 -0.26383056 -0.26383056 -0.16383056 0 -0.26383056
#> 9  09  0.70447928  0.70447928  0.80447928 0  0.70447928
#> 10 10  0.07227781  0.07227781  0.17227781 1  0.17227781

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.60678165 -1.60678165 -1.50678165  -1.3067817  -0.7067817  1  0
#> 2  02  2.30859973  2.30859973  2.40859973   2.6085997   3.2085997  1  0
#> 3  03  0.39798985  0.39798985  0.49798985   0.6979899   1.2979899  0  1
#> 4  04  2.09923734  2.09923734  2.19923734   2.3992373   2.9992373  0  1
#> 5  05  0.38660656  0.38660656  0.48660656   0.6866066   1.2866066  0  1
#> 6  06 -0.05830571 -0.05830571  0.04169429   0.2416943   0.8416943  0  1
#> 7  07  2.04261604  2.04261604  2.14261604   2.3426160   2.9426160  1  0
#> 8  08  0.94329909  0.94329909  1.04329909   1.2432991   1.8432991  0  0
#> 9  09  0.95852940  0.95852940  1.05852940   1.2585294   1.8585294  1  0
#> 10 10  0.06508395  0.06508395  0.16508395   0.3650840   0.9650840  1  1
#>             Y
#> 1  -1.5067817
#> 2   2.4085997
#> 3   0.6979899
#> 4   2.3992373
#> 5   0.6866066
#> 6   0.2416943
#> 7   2.1426160
#> 8   0.9432991
#> 9   1.0585294
#> 10  0.9650840