Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.91510938 -0.91510938 -0.81510938 0 -0.91510938
#> 2  02  0.30736325  0.30736325  0.40736325 0  0.30736325
#> 3  03 -0.88996598 -0.88996598 -0.78996598 1 -0.78996598
#> 4  04  0.69404209  0.69404209  0.79404209 0  0.69404209
#> 5  05  0.07096758  0.07096758  0.17096758 1  0.17096758
#> 6  06  0.22041994  0.22041994  0.32041994 1  0.32041994
#> 7  07  1.60451680  1.60451680  1.70451680 1  1.70451680
#> 8  08 -0.05958493 -0.05958493  0.04041507 0 -0.05958493
#> 9  09 -0.91020866 -0.91020866 -0.81020866 1 -0.81020866
#> 10 10 -2.33387427 -2.33387427 -2.23387427 1 -2.23387427

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.50847051  0.50847051  0.60847051   0.8084705   1.4084705  0  1
#> 2  02 -0.06956048 -0.06956048  0.03043952   0.2304395   0.8304395  1  0
#> 3  03  0.03408425  0.03408425  0.13408425   0.3340842   0.9340842  0  0
#> 4  04 -0.60086975 -0.60086975 -0.50086975  -0.3008697   0.2991303  0  1
#> 5  05 -0.97923910 -0.97923910 -0.87923910  -0.6792391  -0.0792391  0  0
#> 6  06 -1.51673727 -1.51673727 -1.41673727  -1.2167373  -0.6167373  0  1
#> 7  07  1.49084279  1.49084279  1.59084279   1.7908428   2.3908428  1  1
#> 8  08 -0.02473940 -0.02473940  0.07526060   0.2752606   0.8752606  0  0
#> 9  09 -1.05838449 -1.05838449 -0.95838449  -0.7583845  -0.1583845  0  1
#> 10 10  3.07279136  3.07279136  3.17279136   3.3727914   3.9727914  0  0
#>              Y
#> 1   0.80847051
#> 2   0.03043952
#> 3   0.03408425
#> 4  -0.30086975
#> 5  -0.97923910
#> 6  -1.21673727
#> 7   2.39084279
#> 8  -0.02473940
#> 9  -0.75838449
#> 10  3.07279136