Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01  3.0386499  3.0386499  3.1386499 1  3.1386499
#> 2  02 -0.5863653 -0.5863653 -0.4863653 1 -0.4863653
#> 3  03 -0.3541605 -0.3541605 -0.2541605 1 -0.2541605
#> 4  04  0.9639533  0.9639533  1.0639533 1  1.0639533
#> 5  05  1.3327093  1.3327093  1.4327093 0  1.3327093
#> 6  06  0.2548489  0.2548489  0.3548489 1  0.3548489
#> 7  07 -0.8230690 -0.8230690 -0.7230690 0 -0.8230690
#> 8  08  1.4168170  1.4168170  1.5168170 0  1.4168170
#> 9  09  1.7081380  1.7081380  1.8081380 0  1.7081380
#> 10 10 -1.1058387 -1.1058387 -1.0058387 0 -1.1058387

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.21798813  0.21798813  0.31798813  0.51798813  1.11798813  1  0
#> 2  02 -0.91123331 -0.91123331 -0.81123331 -0.61123331 -0.01123331  1  1
#> 3  03  1.09078237  1.09078237  1.19078237  1.39078237  1.99078237  1  1
#> 4  04  0.62512891  0.62512891  0.72512891  0.92512891  1.52512891  0  1
#> 5  05 -0.34616226 -0.34616226 -0.24616226 -0.04616226  0.55383774  0  1
#> 6  06 -0.09764482 -0.09764482  0.00235518  0.20235518  0.80235518  0  0
#> 7  07 -1.35395998 -1.35395998 -1.25395998 -1.05395998 -0.45395998  0  1
#> 8  08  0.18625645  0.18625645  0.28625645  0.48625645  1.08625645  1  1
#> 9  09  0.62635705  0.62635705  0.72635705  0.92635705  1.52635705  1  0
#> 10 10  1.16456314  1.16456314  1.26456314  1.46456314  2.06456314  0  0
#>              Y
#> 1   0.31798813
#> 2  -0.01123331
#> 3   1.99078237
#> 4   0.92512891
#> 5  -0.04616226
#> 6  -0.09764482
#> 7  -1.05395998
#> 8   1.08625645
#> 9   0.72635705
#> 10  1.16456314