Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z           Y
#> 1  01  0.01617189  0.01617189  0.1161719 0  0.01617189
#> 2  02  1.08725115  1.08725115  1.1872511 1  1.18725115
#> 3  03 -0.22622129 -0.22622129 -0.1262213 1 -0.12622129
#> 4  04 -0.67887755 -0.67887755 -0.5788775 0 -0.67887755
#> 5  05  0.31763733  0.31763733  0.4176373 0  0.31763733
#> 6  06  1.27922435  1.27922435  1.3792244 1  1.37922435
#> 7  07 -1.28941589 -1.28941589 -1.1894159 1 -1.18941589
#> 8  08 -1.13581738 -1.13581738 -1.0358174 0 -1.13581738
#> 9  09 -1.50267867 -1.50267867 -1.4026787 0 -1.50267867
#> 10 10  1.06271902  1.06271902  1.1627190 0  1.06271902

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.7861084   0.7861084   0.8861084  1.08610841   1.6861084  0  0
#> 2  02  1.7778806   1.7778806   1.8778806  2.07788064   2.6778806  1  1
#> 3  03  0.1244527   0.1244527   0.2244527  0.42445269   1.0244527  0  0
#> 4  04  1.2171698   1.2171698   1.3171698  1.51716982   2.1171698  1  1
#> 5  05 -0.2558448  -0.2558448  -0.1558448  0.04415521   0.6441552  1  0
#> 6  06 -2.2112837  -2.2112837  -2.1112837 -1.91128374  -1.3112837  0  0
#> 7  07 -1.8594704  -1.8594704  -1.7594704 -1.55947041  -0.9594704  0  1
#> 8  08  0.1864884   0.1864884   0.2864884  0.48648843   1.0864884  1  1
#> 9  09 -1.5101768  -1.5101768  -1.4101768 -1.21017681  -0.6101768  0  0
#> 10 10  0.9916309   0.9916309   1.0916309  1.29163088   1.8916309  0  0
#>             Y
#> 1   0.7861084
#> 2   2.6778806
#> 3   0.1244527
#> 4   2.1171698
#> 5  -0.1558448
#> 6  -2.2112837
#> 7  -1.5594704
#> 8   1.0864884
#> 9  -1.5101768
#> 10  0.9916309