Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID         U     Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -1.602862 -1.602862 -1.5028621 0 -1.6028621
#> 2  02  1.244231  1.244231  1.3442311 1  1.3442311
#> 3  03 -1.086756 -1.086756 -0.9867563 1 -0.9867563
#> 4  04  1.127610  1.127610  1.2276105 0  1.1276105
#> 5  05 -1.383026 -1.383026 -1.2830260 1 -1.2830260
#> 6  06  2.085898  2.085898  2.1858977 0  2.0858977
#> 7  07  1.791692  1.791692  1.8916921 1  1.8916921
#> 8  08 -1.337197 -1.337197 -1.2371970 0 -1.3371970
#> 9  09 -1.561922 -1.561922 -1.4619220 1 -1.4619220
#> 10 10 -1.626927 -1.626927 -1.5269269 0 -1.6269269

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.2959830   0.2959830  0.39598296   0.5959830   1.1959830  1  1
#> 2  02 -1.8967786  -1.8967786 -1.79677865  -1.5967786  -0.9967786  0  0
#> 3  03  1.0814329   1.0814329  1.18143294   1.3814329   1.9814329  0  1
#> 4  04 -1.4162346  -1.4162346 -1.31623461  -1.1162346  -0.5162346  1  1
#> 5  05 -0.1101679  -0.1101679 -0.01016786   0.1898321   0.7898321  0  1
#> 6  06  0.1588804   0.1588804  0.25888035   0.4588804   1.0588804  1  1
#> 7  07  0.7803387   0.7803387  0.88033872   1.0803387   1.6803387  0  0
#> 8  08 -0.6159645  -0.6159645 -0.51596450  -0.3159645   0.2840355  0  0
#> 9  09  2.1386080   2.1386080  2.23860803   2.4386080   3.0386080  0  0
#> 10 10 -0.7052599  -0.7052599 -0.60525994  -0.4052599   0.1947401  0  0
#>             Y
#> 1   1.1959830
#> 2  -1.8967786
#> 3   1.3814329
#> 4  -0.5162346
#> 5   0.1898321
#> 6   1.0588804
#> 7   0.7803387
#> 8  -0.6159645
#> 9   2.1386080
#> 10 -0.7052599