Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.3529221 -0.3529221 -0.25292211 0 -0.35292211
#> 2  02 -0.4518807 -0.4518807 -0.35188067 1 -0.35188067
#> 3  03 -0.1428267 -0.1428267 -0.04282667 1 -0.04282667
#> 4  04  1.2507256  1.2507256  1.35072564 1  1.35072564
#> 5  05 -1.2029758 -1.2029758 -1.10297576 0 -1.20297576
#> 6  06  0.4475176  0.4475176  0.54751759 0  0.44751759
#> 7  07 -0.4874474 -0.4874474 -0.38744737 1 -0.38744737
#> 8  08  1.4035550  1.4035550  1.50355504 1  1.50355504
#> 9  09  0.6026196  0.6026196  0.70261964 1  0.70261964
#> 10 10  1.0648828  1.0648828  1.16488282 0  1.06488282

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  2.1095930   2.1095930   2.2095930   2.4095930  3.00959296  1  1
#> 2  02  0.8763627   0.8763627   0.9763627   1.1763627  1.77636268  1  0
#> 3  03 -0.8185638  -0.8185638  -0.7185638  -0.5185638  0.08143620  1  0
#> 4  04  2.7096179   2.7096179   2.8096179   3.0096179  3.60961791  0  0
#> 5  05 -1.0921889  -1.0921889  -0.9921889  -0.7921889 -0.19218889  0  1
#> 6  06 -1.4200564  -1.4200564  -1.3200564  -1.1200564 -0.52005638  1  1
#> 7  07 -0.9477778  -0.9477778  -0.8477778  -0.6477778 -0.04777775  1  1
#> 8  08 -1.1977639  -1.1977639  -1.0977639  -0.8977639 -0.29776393  1  1
#> 9  09 -1.1870169  -1.1870169  -1.0870169  -0.8870169 -0.28701694  1  0
#> 10 10  1.3322168   1.3322168   1.4322168   1.6322168  2.23221675  1  1
#>              Y
#> 1   3.00959296
#> 2   0.97636268
#> 3  -0.71856380
#> 4   2.70961791
#> 5  -0.79218889
#> 6  -0.52005638
#> 7  -0.04777775
#> 8  -0.29776393
#> 9  -1.08701694
#> 10  2.23221675