Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -0.6607790 -0.6607790 -0.5607790 1 -0.5607790
#> 2  02  1.0502049  1.0502049  1.1502049 1  1.1502049
#> 3  03 -0.9085622 -0.9085622 -0.8085622 1 -0.8085622
#> 4  04  0.9767747  0.9767747  1.0767747 1  1.0767747
#> 5  05  1.1152828  1.1152828  1.2152828 0  1.1152828
#> 6  06  0.1162778  0.1162778  0.2162778 0  0.1162778
#> 7  07 -0.2863918 -0.2863918 -0.1863918 1 -0.1863918
#> 8  08 -1.0649143 -1.0649143 -0.9649143 1 -0.9649143
#> 9  09  0.1626896  0.1626896  0.2626896 1  0.2626896
#> 10 10  1.1814235  1.1814235  1.2814235 0  1.1814235

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.40748660  1.40748660   1.5074866  1.70748660   2.3074866  1  1
#> 2  02 -0.27286999 -0.27286999  -0.1728700  0.02713001   0.6271300  0  1
#> 3  03 -1.99355749 -1.99355749  -1.8935575 -1.69355749  -1.0935575  0  0
#> 4  04  1.42844396  1.42844396   1.5284440  1.72844396   2.3284440  0  1
#> 5  05  0.09924630  0.09924630   0.1992463  0.39924630   0.9992463  1  1
#> 6  06 -1.75359524 -1.75359524  -1.6535952 -1.45359524  -0.8535952  0  1
#> 7  07 -0.62163453 -0.62163453  -0.5216345 -0.32163453   0.2783655  1  0
#> 8  08  0.05616494  0.05616494   0.1561649  0.35616494   0.9561649  0  0
#> 9  09  0.95628930  0.95628930   1.0562893  1.25628930   1.8562893  0  1
#> 10 10 -0.64608689 -0.64608689  -0.5460869 -0.34608689   0.2539131  0  0
#>              Y
#> 1   2.30748660
#> 2   0.02713001
#> 3  -1.99355749
#> 4   1.72844396
#> 5   0.99924630
#> 6  -1.45359524
#> 7  -0.52163453
#> 8   0.05616494
#> 9   1.25628930
#> 10 -0.64608689