Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -0.70267144 -0.70267144 -0.6026714 0 -0.7026714
#> 2  02 -2.12157963 -2.12157963 -2.0215796 0 -2.1215796
#> 3  03 -0.41825629 -0.41825629 -0.3182563 1 -0.3182563
#> 4  04  0.42846877  0.42846877  0.5284688 1  0.5284688
#> 5  05  0.07619132  0.07619132  0.1761913 1  0.1761913
#> 6  06 -1.51395865 -1.51395865 -1.4139586 1 -1.4139586
#> 7  07  0.38699660  0.38699660  0.4869966 0  0.3869966
#> 8  08  0.52617473  0.52617473  0.6261747 1  0.6261747
#> 9  09 -1.13000339 -1.13000339 -1.0300034 0 -1.1300034
#> 10 10 -0.83467227 -0.83467227 -0.7346723 1 -0.7346723

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.47436873 -0.47436873 -0.37436873  -0.1743687  0.425631269  0  1
#> 2  02 -0.90905950 -0.90905950 -0.80905950  -0.6090595 -0.009059496  0  1
#> 3  03  0.87716002  0.87716002  0.97716002   1.1771600  1.777160016  1  1
#> 4  04  0.71796980  0.71796980  0.81796980   1.0179698  1.617969797  1  1
#> 5  05 -0.06086011 -0.06086011  0.03913989   0.2391399  0.839139894  0  0
#> 6  06 -0.88355165 -0.88355165 -0.78355165  -0.5835516  0.016448351  0  0
#> 7  07  0.78008020  0.78008020  0.88008020   1.0800802  1.680080203  1  0
#> 8  08 -0.06357471 -0.06357471  0.03642529   0.2364253  0.836425293  0  1
#> 9  09 -1.43166067 -1.43166067 -1.33166067  -1.1316607 -0.531660668  0  0
#> 10 10  0.13691847  0.13691847  0.23691847   0.4369185  1.036918471  0  0
#>              Y
#> 1  -0.17436873
#> 2  -0.60905950
#> 3   1.77716002
#> 4   1.61796980
#> 5  -0.06086011
#> 6  -0.88355165
#> 7   0.88008020
#> 8   0.23642529
#> 9  -1.43166067
#> 10  0.13691847