Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.14023644  0.14023644  0.24023644 0  0.14023644
#> 2  02 -1.20545200 -1.20545200 -1.10545200 0 -1.20545200
#> 3  03  0.83660373  0.83660373  0.93660373 1  0.93660373
#> 4  04 -1.42296286 -1.42296286 -1.32296286 1 -1.32296286
#> 5  05  0.56984355  0.56984355  0.66984355 1  0.66984355
#> 6  06 -0.03942656 -0.03942656  0.06057344 1  0.06057344
#> 7  07  0.19670612  0.19670612  0.29670612 0  0.19670612
#> 8  08  0.20956426  0.20956426  0.30956426 0  0.20956426
#> 9  09  0.43425120  0.43425120  0.53425120 0  0.43425120
#> 10 10  0.76146449  0.76146449  0.86146449 1  0.86146449

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.1754463   0.1754463  0.27544625  0.47544625  1.07544625  0  0
#> 2  02  0.3123309   0.3123309  0.41233089  0.61233089  1.21233089  0  0
#> 3  03 -0.3975928  -0.3975928 -0.29759285 -0.09759285  0.50240715  1  1
#> 4  04 -0.8765132  -0.8765132 -0.77651324 -0.57651324  0.02348676  1  0
#> 5  05  0.5361340   0.5361340  0.63613396  0.83613396  1.43613396  1  1
#> 6  06  0.8727449   0.8727449  0.97274492  1.17274492  1.77274492  1  0
#> 7  07 -0.1273770  -0.1273770 -0.02737703  0.17262297  0.77262297  0  0
#> 8  08 -0.1569656  -0.1569656 -0.05696561  0.14303439  0.74303439  0  1
#> 9  09 -0.6453816  -0.6453816 -0.54538157 -0.34538157  0.25461843  0  0
#> 10 10  1.0044984   1.0044984  1.10449837  1.30449837  1.90449837  1  0
#>             Y
#> 1   0.1754463
#> 2   0.3123309
#> 3   0.5024072
#> 4  -0.7765132
#> 5   1.4361340
#> 6   0.9727449
#> 7  -0.1273770
#> 8   0.1430344
#> 9  -0.6453816
#> 10  1.1044984