Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z            Y
#> 1  01 -0.09110741 -0.09110741  0.008892594 1  0.008892594
#> 2  02 -1.03248493 -1.03248493 -0.932484932 0 -1.032484932
#> 3  03 -0.02337475 -0.02337475  0.076625248 0 -0.023374752
#> 4  04  0.57413883  0.57413883  0.674138826 0  0.574138826
#> 5  05  1.77066858  1.77066858  1.870668579 0  1.770668579
#> 6  06  1.19067302  1.19067302  1.290673015 1  1.290673015
#> 7  07 -0.01727228 -0.01727228  0.082727716 1  0.082727716
#> 8  08  1.44195616  1.44195616  1.541956157 0  1.441956157
#> 9  09 -0.85992562 -0.85992562 -0.759925620 1 -0.759925620
#> 10 10  1.21617857  1.21617857  1.316178569 0  1.216178569

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.20905887 -0.20905887 -0.10905887  0.09094113   0.6909411  0  1
#> 2  02  0.42861590  0.42861590  0.52861590  0.72861590   1.3286159  0  1
#> 3  03 -0.04681421 -0.04681421  0.05318579  0.25318579   0.8531858  1  1
#> 4  04 -1.20092805 -1.20092805 -1.10092805 -0.90092805  -0.3009280  0  0
#> 5  05  1.09800071  1.09800071  1.19800071  1.39800071   1.9980007  1  1
#> 6  06  1.00683740  1.00683740  1.10683740  1.30683740   1.9068374  0  1
#> 7  07  0.04926025  0.04926025  0.14926025  0.34926025   0.9492603  1  0
#> 8  08 -1.09826595 -1.09826595 -0.99826595 -0.79826595  -0.1982659  1  0
#> 9  09  1.04291684  1.04291684  1.14291684  1.34291684   1.9429168  1  0
#> 10 10  0.64292888  0.64292888  0.74292888  0.94292888   1.5429289  1  0
#>              Y
#> 1   0.09094113
#> 2   0.72861590
#> 3   0.85318579
#> 4  -1.20092805
#> 5   1.99800071
#> 6   1.30683740
#> 7   0.14926025
#> 8  -0.99826595
#> 9   1.14291684
#> 10  0.74292888