Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  1.76677947  1.76677947  1.86677947 1  1.86677947
#> 2  02  0.90163511  0.90163511  1.00163511 1  1.00163511
#> 3  03  0.69557857  0.69557857  0.79557857 1  0.79557857
#> 4  04 -0.03416081 -0.03416081  0.06583919 0 -0.03416081
#> 5  05  1.16334578  1.16334578  1.26334578 0  1.16334578
#> 6  06  0.36887207  0.36887207  0.46887207 1  0.46887207
#> 7  07  0.98737658  0.98737658  1.08737658 1  1.08737658
#> 8  08  0.16589021  0.16589021  0.26589021 0  0.16589021
#> 9  09 -0.46651285 -0.46651285 -0.36651285 1 -0.36651285
#> 10 10  0.57684850  0.57684850  0.67684850 1  0.67684850

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.82683029  0.82683029  0.92683029   1.1268303  1.726830294  0  1
#> 2  02  1.18598816  1.18598816  1.28598816   1.4859882  2.085988161  0  1
#> 3  03 -0.03905872 -0.03905872  0.06094128   0.2609413  0.860941276  0  0
#> 4  04 -0.01837184 -0.01837184  0.08162816   0.2816282  0.881628162  0  0
#> 5  05  1.62647370  1.62647370  1.72647370   1.9264737  2.526473704  1  1
#> 6  06 -0.07803476 -0.07803476  0.02196524   0.2219652  0.821965239  0  0
#> 7  07  1.27614412  1.27614412  1.37614412   1.5761441  2.176144116  1  1
#> 8  08  0.93744168  0.93744168  1.03744168   1.2374417  1.837441676  0  0
#> 9  09  0.72142260  0.72142260  0.82142260   1.0214226  1.621422597  0  0
#> 10 10 -0.90839268 -0.90839268 -0.80839268  -0.6083927 -0.008392679  1  0
#>              Y
#> 1   1.12683029
#> 2   1.48598816
#> 3  -0.03905872
#> 4  -0.01837184
#> 5   2.52647370
#> 6  -0.07803476
#> 7   2.17614412
#> 8   0.93744168
#> 9   0.72142260
#> 10 -0.80839268