Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01  1.0305529  1.0305529  1.1305529 1  1.1305529
#> 2  02  0.9383468  0.9383468  1.0383468 0  0.9383468
#> 3  03 -0.8166572 -0.8166572 -0.7166572 0 -0.8166572
#> 4  04  1.9881283  1.9881283  2.0881283 1  2.0881283
#> 5  05  0.8965980  0.8965980  0.9965980 1  0.9965980
#> 6  06  1.1654485  1.1654485  1.2654485 1  1.2654485
#> 7  07 -1.6656247 -1.6656247 -1.5656247 1 -1.5656247
#> 8  08  1.9829734  1.9829734  2.0829734 1  2.0829734
#> 9  09 -0.6394197 -0.6394197 -0.5394197 1 -0.5394197
#> 10 10 -1.0592701 -1.0592701 -0.9592701 1 -0.9592701

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.49969177  1.49969177  1.59969177   1.7996918   2.3996918  1  1
#> 2  02 -0.35707630 -0.35707630 -0.25707630  -0.0570763   0.5429237  1  0
#> 3  03  0.18663013  0.18663013  0.28663013   0.4866301   1.0866301  0  0
#> 4  04 -0.11331489 -0.11331489 -0.01331489   0.1866851   0.7866851  1  1
#> 5  05  1.62590309  1.62590309  1.72590309   1.9259031   2.5259031  1  0
#> 6  06 -0.18888614 -0.18888614 -0.08888614   0.1111139   0.7111139  1  1
#> 7  07  0.80353013  0.80353013  0.90353013   1.1035301   1.7035301  1  1
#> 8  08  0.24922680  0.24922680  0.34922680   0.5492268   1.1492268  0  0
#> 9  09 -0.07030232 -0.07030232  0.02969768   0.2296977   0.8296977  0  0
#> 10 10  1.24042958  1.24042958  1.34042958   1.5404296   2.1404296  1  0
#>              Y
#> 1   2.39969177
#> 2  -0.25707630
#> 3   0.18663013
#> 4   0.78668511
#> 5   1.72590309
#> 6   0.71111386
#> 7   1.70353013
#> 8   0.24922680
#> 9  -0.07030232
#> 10  1.34042958