Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z          Y
#> 1  01 -0.1578506 -0.1578506 -0.05785058 0 -0.1578506
#> 2  02  0.1316356  0.1316356  0.23163561 1  0.2316356
#> 3  03 -1.0785917 -1.0785917 -0.97859167 0 -1.0785917
#> 4  04  2.1822126  2.1822126  2.28221258 0  2.1822126
#> 5  05  0.4183623  0.4183623  0.51836231 1  0.5183623
#> 6  06 -1.2122600 -1.2122600 -1.11226004 1 -1.1122600
#> 7  07 -0.9321757 -0.9321757 -0.83217572 0 -0.9321757
#> 8  08 -0.4093773 -0.4093773 -0.30937733 1 -0.3093773
#> 9  09  0.6303799  0.6303799  0.73037992 0  0.6303799
#> 10 10  0.7407389  0.7407389  0.84073889 0  0.7407389

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.42769134 -0.42769134 -0.32769134 -0.12769134  0.47230866  0  0
#> 2  02 -2.45009808 -2.45009808 -2.35009808 -2.15009808 -1.55009808  1  0
#> 3  03 -2.13892288 -2.13892288 -2.03892288 -1.83892288 -1.23892288  0  0
#> 4  04 -0.98435338 -0.98435338 -0.88435338 -0.68435338 -0.08435338  0  0
#> 5  05 -0.20638609 -0.20638609 -0.10638609  0.09361391  0.69361391  0  0
#> 6  06 -0.04331578 -0.04331578  0.05668422  0.25668422  0.85668422  1  1
#> 7  07 -2.59242006 -2.59242006 -2.49242006 -2.29242006 -1.69242006  0  0
#> 8  08 -0.62574680 -0.62574680 -0.52574680 -0.32574680  0.27425320  0  1
#> 9  09  0.78032901  0.78032901  0.88032901  1.08032901  1.68032901  1  1
#> 10 10  0.13467733  0.13467733  0.23467733  0.43467733  1.03467733  1  0
#>             Y
#> 1  -0.4276913
#> 2  -2.3500981
#> 3  -2.1389229
#> 4  -0.9843534
#> 5  -0.2063861
#> 6   0.8566842
#> 7  -2.5924201
#> 8  -0.3257468
#> 9   1.6803290
#> 10  0.2346773