Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -0.2823675 -0.2823675 -0.1823675 1 -0.1823675
#> 2  02 -0.8367524 -0.8367524 -0.7367524 1 -0.7367524
#> 3  03  0.7575772  0.7575772  0.8575772 0  0.7575772
#> 4  04 -1.3762185 -1.3762185 -1.2762185 1 -1.2762185
#> 5  05  0.4327603  0.4327603  0.5327603 0  0.4327603
#> 6  06  1.5889067  1.5889067  1.6889067 0  1.5889067
#> 7  07  1.2484613  1.2484613  1.3484613 0  1.2484613
#> 8  08 -0.5058432 -0.5058432 -0.4058432 0 -0.5058432
#> 9  09  0.2948407  0.2948407  0.3948407 1  0.3948407
#> 10 10  1.1367001  1.1367001  1.2367001 0  1.1367001

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0  Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.2549592   1.2549592  1.354959161   1.5549592   2.1549592  1  1
#> 2  02 -0.1050153  -0.1050153 -0.005015290   0.1949847   0.7949847  1  1
#> 3  03  0.1912558   0.1912558  0.291255780   0.4912558   1.0912558  0  1
#> 4  04  0.3382956   0.3382956  0.438295604   0.6382956   1.2382956  1  1
#> 5  05  0.0733859   0.0733859  0.173385899   0.3733859   0.9733859  1  0
#> 6  06  2.2261209   2.2261209  2.326120866   2.5261209   3.1261209  1  1
#> 7  07  0.7517762   0.7517762  0.851776229   1.0517762   1.6517762  1  0
#> 8  08  0.3571870   0.3571870  0.457187033   0.6571870   1.2571870  1  1
#> 9  09  0.2735069   0.2735069  0.373506931   0.5735069   1.1735069  0  0
#> 10 10 -0.1023437  -0.1023437 -0.002343739   0.1976563   0.7976563  1  0
#>               Y
#> 1   2.154959161
#> 2   0.794984710
#> 3   0.491255780
#> 4   1.238295604
#> 5   0.173385899
#> 6   3.126120866
#> 7   0.851776229
#> 8   1.257187033
#> 9   0.273506931
#> 10 -0.002343739