Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.5243243  0.5243243  0.62432425 1  0.62432425
#> 2  02  0.8462750  0.8462750  0.94627499 0  0.84627499
#> 3  03 -0.1716279 -0.1716279 -0.07162787 1 -0.07162787
#> 4  04  0.6166723  0.6166723  0.71667234 0  0.61667234
#> 5  05 -0.4413598 -0.4413598 -0.34135982 0 -0.44135982
#> 6  06 -0.9802502 -0.9802502 -0.88025023 0 -0.98025023
#> 7  07 -0.4780406 -0.4780406 -0.37804064 1 -0.37804064
#> 8  08 -0.5271182 -0.5271182 -0.42711815 0 -0.52711815
#> 9  09  0.7568381  0.7568381  0.85683806 0  0.75683806
#> 10 10  2.3349772  2.3349772  2.43497721 1  2.43497721

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.9071220  -0.9071220 -0.80712205 -0.60712205 -0.00712205  1  0
#> 2  02 -1.3254571  -1.3254571 -1.22545715 -1.02545715 -0.42545715  1  0
#> 3  03  0.6201268   0.6201268  0.72012678  0.92012678  1.52012678  1  1
#> 4  04  1.0593535   1.0593535  1.15935351  1.35935351  1.95935351  1  1
#> 5  05 -0.1876241  -0.1876241 -0.08762408  0.11237592  0.71237592  0  0
#> 6  06  0.2214557   0.2214557  0.32145566  0.52145566  1.12145566  1  0
#> 7  07  1.4635113   1.4635113  1.56351134  1.76351134  2.36351134  0  1
#> 8  08 -1.4741652  -1.4741652 -1.37416524 -1.17416524 -0.57416524  0  1
#> 9  09 -0.2199709  -0.2199709 -0.11997093  0.08002907  0.68002907  1  0
#> 10 10  1.0885575   1.0885575  1.18855750  1.38855750  1.98855750  0  1
#>             Y
#> 1  -0.8071220
#> 2  -1.2254571
#> 3   1.5201268
#> 4   1.9593535
#> 5  -0.1876241
#> 6   0.3214557
#> 7   1.7635113
#> 8  -1.1741652
#> 9  -0.1199709
#> 10  1.3885575