Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z           Y
#> 1  01  0.04485278  0.04485278  0.1448528 0  0.04485278
#> 2  02 -0.31386149 -0.31386149 -0.2138615 1 -0.21386149
#> 3  03  1.70911318  1.70911318  1.8091132 0  1.70911318
#> 4  04  0.47732444  0.47732444  0.5773244 1  0.57732444
#> 5  05  0.50390727  0.50390727  0.6039073 0  0.50390727
#> 6  06 -0.38416185 -0.38416185 -0.2841619 1 -0.28416185
#> 7  07 -0.55407996 -0.55407996 -0.4540800 0 -0.55407996
#> 8  08 -0.34525076 -0.34525076 -0.2452508 0 -0.34525076
#> 9  09  0.36738418  0.36738418  0.4673842 1  0.46738418
#> 10 10  0.51066707  0.51066707  0.6106671 1  0.61066707

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.05792605 -1.05792605 -0.95792605  -0.7579260  -0.1579260  0  1
#> 2  02  1.50682244  1.50682244  1.60682244   1.8068224   2.4068224  0  1
#> 3  03 -1.75413014 -1.75413014 -1.65413014  -1.4541301  -0.8541301  1  0
#> 4  04 -0.56492522 -0.56492522 -0.46492522  -0.2649252   0.3350748  0  0
#> 5  05  0.26523687  0.26523687  0.36523687   0.5652369   1.1652369  1  1
#> 6  06  1.63137754  1.63137754  1.73137754   1.9313775   2.5313775  1  0
#> 7  07  0.16104536  0.16104536  0.26104536   0.4610454   1.0610454  0  0
#> 8  08 -0.51747844 -0.51747844 -0.41747844  -0.2174784   0.3825216  0  1
#> 9  09 -0.04138152 -0.04138152  0.05861848   0.2586185   0.8586185  0  0
#> 10 10  0.57237620  0.57237620  0.67237620   0.8723762   1.4723762  0  0
#>              Y
#> 1  -0.75792605
#> 2   1.80682244
#> 3  -1.65413014
#> 4  -0.56492522
#> 5   1.16523687
#> 6   1.73137754
#> 7   0.16104536
#> 8  -0.21747844
#> 9  -0.04138152
#> 10  0.57237620