Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z            Y
#> 1  01  0.30694134  0.30694134  0.406941339 0  0.306941339
#> 2  02 -1.11342838 -1.11342838 -1.013428383 1 -1.013428383
#> 3  03 -0.48091530 -0.48091530 -0.380915299 1 -0.380915299
#> 4  04  1.30177312  1.30177312  1.401773123 0  1.301773123
#> 5  05  0.51793043  0.51793043  0.617930431 0  0.517930431
#> 6  06  2.76380930  2.76380930  2.863809295 1  2.863809295
#> 7  07  0.26278796  0.26278796  0.362787962 1  0.362787962
#> 8  08  0.79596004  0.79596004  0.895960042 0  0.795960042
#> 9  09  0.82248009  0.82248009  0.922480088 1  0.922480088
#> 10 10 -0.09043279 -0.09043279  0.009567211 1  0.009567211

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.60492323  0.60492323   0.7049232   0.9049232   1.5049232  1  1
#> 2  02  1.52002704  1.52002704   1.6200270   1.8200270   2.4200270  0  0
#> 3  03  1.32476326  1.32476326   1.4247633   1.6247633   2.2247633  0  0
#> 4  04 -1.11384602 -1.11384602  -1.0138460  -0.8138460  -0.2138460  0  1
#> 5  05  1.28999502  1.28999502   1.3899950   1.5899950   2.1899950  0  1
#> 6  06 -1.39526341 -1.39526341  -1.2952634  -1.0952634  -0.4952634  1  0
#> 7  07 -1.16502787 -1.16502787  -1.0650279  -0.8650279  -0.2650279  1  0
#> 8  08  1.09184482  1.09184482   1.1918448   1.3918448   1.9918448  0  1
#> 9  09  1.04310875  1.04310875   1.1431088   1.3431088   1.9431088  1  1
#> 10 10  0.07871566  0.07871566   0.1787157   0.3787157   0.9787157  0  0
#>              Y
#> 1   1.50492323
#> 2   1.52002704
#> 3   1.32476326
#> 4  -0.81384602
#> 5   1.58999502
#> 6  -1.29526341
#> 7  -1.06502787
#> 8   1.39184482
#> 9   1.94310875
#> 10  0.07871566