Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.68376037 -0.68376037 -0.58376037 0 -0.68376037
#> 2  02  0.48825985  0.48825985  0.58825985 0  0.48825985
#> 3  03 -0.08270712 -0.08270712  0.01729288 1  0.01729288
#> 4  04 -0.33382803 -0.33382803 -0.23382803 0 -0.33382803
#> 5  05  0.44265454  0.44265454  0.54265454 0  0.44265454
#> 6  06 -2.51003377 -2.51003377 -2.41003377 1 -2.41003377
#> 7  07  0.18383886  0.18383886  0.28383886 0  0.18383886
#> 8  08  0.73502783  0.73502783  0.83502783 0  0.73502783
#> 9  09  0.61265090  0.61265090  0.71265090 0  0.61265090
#> 10 10 -0.06543338 -0.06543338  0.03456662 1  0.03456662

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.2297260   0.2297260   0.3297260   0.5297260   1.1297260  0  0
#> 2  02 -0.7381505  -0.7381505  -0.6381505  -0.4381505   0.1618495  1  0
#> 3  03 -0.5211016  -0.5211016  -0.4211016  -0.2211016   0.3788984  1  0
#> 4  04  0.7745857   0.7745857   0.8745857   1.0745857   1.6745857  1  1
#> 5  05  1.0077211   1.0077211   1.1077211   1.3077211   1.9077211  1  1
#> 6  06 -0.9828514  -0.9828514  -0.8828514  -0.6828514  -0.0828514  1  1
#> 7  07 -0.5264846  -0.5264846  -0.4264846  -0.2264846   0.3735154  0  0
#> 8  08  1.8127204   1.8127204   1.9127204   2.1127204   2.7127204  0  1
#> 9  09 -1.8795438  -1.8795438  -1.7795438  -1.5795438  -0.9795438  0  1
#> 10 10 -1.2429061  -1.2429061  -1.1429061  -0.9429061  -0.3429061  0  1
#>             Y
#> 1   0.2297260
#> 2  -0.6381505
#> 3  -0.4211016
#> 4   1.6745857
#> 5   1.9077211
#> 6  -0.0828514
#> 7  -0.5264846
#> 8   2.1127204
#> 9  -1.5795438
#> 10 -0.9429061