Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.2814955  0.2814955  0.38149553 0  0.28149553
#> 2  02 -0.1181434 -0.1181434 -0.01814342 0 -0.11814342
#> 3  03  1.4320714  1.4320714  1.53207139 0  1.43207139
#> 4  04 -1.5781811 -1.5781811 -1.47818110 0 -1.57818110
#> 5  05  0.7178007  0.7178007  0.81780065 1  0.81780065
#> 6  06 -0.6730134 -0.6730134 -0.57301340 1 -0.57301340
#> 7  07 -0.6376508 -0.6376508 -0.53765078 1 -0.53765078
#> 8  08 -0.5496671 -0.5496671 -0.44966714 0 -0.54966714
#> 9  09  0.2264086  0.2264086  0.32640858 1  0.32640858
#> 10 10 -0.1274307 -0.1274307 -0.02743068 1 -0.02743068

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.04942978 -1.04942978 -0.94942978 -0.74942978  -0.1494298  0  0
#> 2  02  0.18363658  0.18363658  0.28363658  0.48363658   1.0836366  0  1
#> 3  03 -0.51092256 -0.51092256 -0.41092256 -0.21092256   0.3890774  1  1
#> 4  04  0.33513721  0.33513721  0.43513721  0.63513721   1.2351372  0  0
#> 5  05 -0.08182491 -0.08182491  0.01817509  0.21817509   0.8181751  0  1
#> 6  06  2.31130622  2.31130622  2.41130622  2.61130622   3.2113062  0  1
#> 7  07  0.05808061  0.05808061  0.15808061  0.35808061   0.9580806  1  0
#> 8  08 -0.32448859 -0.32448859 -0.22448859 -0.02448859   0.5755114  1  0
#> 9  09  0.06750985  0.06750985  0.16750985  0.36750985   0.9675099  0  0
#> 10 10  0.48236437  0.48236437  0.58236437  0.78236437   1.3823644  0  0
#>              Y
#> 1  -1.04942978
#> 2   0.48363658
#> 3   0.38907744
#> 4   0.33513721
#> 5   0.21817509
#> 6   2.61130622
#> 7   0.15808061
#> 8  -0.22448859
#> 9   0.06750985
#> 10  0.48236437