Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -1.46667818 -1.46667818 -1.36667818 1 -1.36667818
#> 2  02 -0.01190972 -0.01190972  0.08809028 0 -0.01190972
#> 3  03 -0.56870536 -0.56870536 -0.46870536 0 -0.56870536
#> 4  04  0.09283712  0.09283712  0.19283712 0  0.09283712
#> 5  05  0.66611102  0.66611102  0.76611102 1  0.76611102
#> 6  06 -0.55778832 -0.55778832 -0.45778832 0 -0.55778832
#> 7  07 -0.48463583 -0.48463583 -0.38463583 1 -0.38463583
#> 8  08 -0.75236355 -0.75236355 -0.65236355 1 -0.65236355
#> 9  09  0.61524764  0.61524764  0.71524764 0  0.61524764
#> 10 10  0.90920459  0.90920459  1.00920459 1  1.00920459

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1  Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.9065873  -0.9065873  -0.8065873 -0.60658735 -0.006587346  0  0
#> 2  02  0.3188664   0.3188664   0.4188664  0.61886642  1.218866419  1  1
#> 3  03 -1.0480759  -1.0480759  -0.9480759 -0.74807587 -0.148075870  1  1
#> 4  04  0.2001062   0.2001062   0.3001062  0.50010623  1.100106231  0  1
#> 5  05 -1.5077531  -1.5077531  -1.4077531 -1.20775309 -0.607753093  1  1
#> 6  06 -0.8847778  -0.8847778  -0.7847778 -0.58477776  0.015222238  1  1
#> 7  07 -1.2769589  -1.2769589  -1.1769589 -0.97695888 -0.376958877  1  0
#> 8  08  0.7282625   0.7282625   0.8282625  1.02826253  1.628262534  0  1
#> 9  09  0.2180865   0.2180865   0.3180865  0.51808652  1.118086525  0  1
#> 10 10 -0.3927890  -0.3927890  -0.2927890 -0.09278898  0.507211025  0  1
#>              Y
#> 1  -0.90658735
#> 2   1.21886642
#> 3  -0.14807587
#> 4   0.50010623
#> 5  -0.60775309
#> 6   0.01522224
#> 7  -1.17695888
#> 8   1.02826253
#> 9   0.51808652
#> 10 -0.09278898