Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.97893459  0.97893459  1.07893459 1  1.07893459
#> 2  02  0.52148384  0.52148384  0.62148384 0  0.52148384
#> 3  03 -1.05353318 -1.05353318 -0.95353318 1 -0.95353318
#> 4  04 -1.60409039 -1.60409039 -1.50409039 0 -1.60409039
#> 5  05  1.18821761  1.18821761  1.28821761 1  1.28821761
#> 6  06  0.62008333  0.62008333  0.72008333 1  0.72008333
#> 7  07  0.97915562  0.97915562  1.07915562 1  1.07915562
#> 8  08 -0.13292110 -0.13292110 -0.03292110 0 -0.13292110
#> 9  09 -0.01223588 -0.01223588  0.08776412 0 -0.01223588
#> 10 10 -0.69645361 -0.69645361 -0.59645361 1 -0.59645361

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.44953640  0.44953640  0.54953640  0.74953640   1.3495364  1  0
#> 2  02  0.94635049  0.94635049  1.04635049  1.24635049   1.8463505  0  0
#> 3  03  1.16778807  1.16778807  1.26778807  1.46778807   2.0677881  1  0
#> 4  04 -0.01158343 -0.01158343  0.08841657  0.28841657   0.8884166  1  1
#> 5  05  0.92614999  0.92614999  1.02614999  1.22614999   1.8261500  0  0
#> 6  06 -0.28767291 -0.28767291 -0.18767291  0.01232709   0.6123271  1  1
#> 7  07  1.91485042  1.91485042  2.01485042  2.21485042   2.8148504  1  0
#> 8  08  1.25957558  1.25957558  1.35957558  1.55957558   2.1595756  0  1
#> 9  09 -0.33687623 -0.33687623 -0.23687623 -0.03687623   0.5631238  1  0
#> 10 10 -0.58799694 -0.58799694 -0.48799694 -0.28799694   0.3120031  1  0
#>             Y
#> 1   0.5495364
#> 2   0.9463505
#> 3   1.2677881
#> 4   0.8884166
#> 5   0.9261500
#> 6   0.6123271
#> 7   2.0148504
#> 8   1.5595756
#> 9  -0.2368762
#> 10 -0.4879969