Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z          Y
#> 1  01  0.72729013  0.72729013  0.82729013 1  0.8272901
#> 2  02 -1.50664957 -1.50664957 -1.40664957 0 -1.5066496
#> 3  03 -0.23604320 -0.23604320 -0.13604320 0 -0.2360432
#> 4  04  0.32379209  0.32379209  0.42379209 0  0.3237921
#> 5  05  0.42827341  0.42827341  0.52827341 1  0.5282734
#> 6  06 -0.94130287 -0.94130287 -0.84130287 1 -0.8413029
#> 7  07  0.27469399  0.27469399  0.37469399 1  0.3746940
#> 8  08 -0.11056181 -0.11056181 -0.01056181 0 -0.1105618
#> 9  09  0.02499551  0.02499551  0.12499551 1  0.1249955
#> 10 10 -0.13574780 -0.13574780 -0.03574780 0 -0.1357478

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.5684433   1.5684433   1.6684433   1.8684433   2.4684433  1  0
#> 2  02  0.6785968   0.6785968   0.7785968   0.9785968   1.5785968  1  0
#> 3  03  2.9384520   2.9384520   3.0384520   3.2384520   3.8384520  0  0
#> 4  04  0.9684637   0.9684637   1.0684637   1.2684637   1.8684637  0  1
#> 5  05 -1.0889764  -1.0889764  -0.9889764  -0.7889764  -0.1889764  1  1
#> 6  06 -0.4562559  -0.4562559  -0.3562559  -0.1562559   0.4437441  1  1
#> 7  07  2.6062666   2.6062666   2.7062666   2.9062666   3.5062666  0  0
#> 8  08  0.5949755   0.5949755   0.6949755   0.8949755   1.4949755  1  1
#> 9  09  0.4905864   0.4905864   0.5905864   0.7905864   1.3905864  1  0
#> 10 10  0.8095457   0.8095457   0.9095457   1.1095457   1.7095457  1  0
#>             Y
#> 1   1.6684433
#> 2   0.7785968
#> 3   2.9384520
#> 4   1.2684637
#> 5  -0.1889764
#> 6   0.4437441
#> 7   2.6062666
#> 8   1.4949755
#> 9   0.5905864
#> 10  0.9095457