Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z            Y
#> 1  01  0.03245864  0.03245864  0.132458636 0  0.032458636
#> 2  02 -0.29335472 -0.29335472 -0.193354723 0 -0.293354723
#> 3  03  0.70856919  0.70856919  0.808569195 0  0.708569195
#> 4  04  0.85948376  0.85948376  0.959483756 0  0.859483756
#> 5  05 -1.18538224 -1.18538224 -1.085382240 0 -1.185382240
#> 6  06  1.16011383  1.16011383  1.260113827 0  1.160113827
#> 7  07  0.76972753  0.76972753  0.869727527 1  0.869727527
#> 8  08 -0.66246464 -0.66246464 -0.562464635 0 -0.662464635
#> 9  09 -0.09083554 -0.09083554  0.009164463 1  0.009164463
#> 10 10 -1.21595789 -1.21595789 -1.115957885 0 -1.215957885

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.57677873  0.57677873   0.6767787  0.87677873   1.4767787  1  1
#> 2  02  0.06433143  0.06433143   0.1643314  0.36433143   0.9643314  0  0
#> 3  03 -0.32032386 -0.32032386  -0.2203239 -0.02032386   0.5796761  1  1
#> 4  04 -1.57448764 -1.57448764  -1.4744876 -1.27448764  -0.6744876  1  1
#> 5  05 -0.44897925 -0.44897925  -0.3489792 -0.14897925   0.4510208  0  0
#> 6  06 -0.20613393 -0.20613393  -0.1061339  0.09386607   0.6938661  0  1
#> 7  07  2.37573722  2.37573722   2.4757372  2.67573722   3.2757372  0  1
#> 8  08 -0.18885090 -0.18885090  -0.0888509  0.11114910   0.7111491  1  0
#> 9  09 -1.17670224 -1.17670224  -1.0767022 -0.87670224  -0.2767022  1  0
#> 10 10 -1.59840563 -1.59840563  -1.4984056 -1.29840563  -0.6984056  1  0
#>              Y
#> 1   1.47677873
#> 2   0.06433143
#> 3   0.57967614
#> 4  -0.67448764
#> 5  -0.44897925
#> 6   0.09386607
#> 7   2.67573722
#> 8  -0.08885090
#> 9  -1.07670224
#> 10 -1.49840563