Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01  1.0463515  1.0463515  1.1463515 1  1.1463515
#> 2  02 -0.4687186 -0.4687186 -0.3687186 1 -0.3687186
#> 3  03 -0.3330535 -0.3330535 -0.2330535 0 -0.3330535
#> 4  04 -1.9259610 -1.9259610 -1.8259610 1 -1.8259610
#> 5  05 -0.8438076 -0.8438076 -0.7438076 0 -0.8438076
#> 6  06  1.4993099  1.4993099  1.5993099 1  1.5993099
#> 7  07  0.5447141  0.5447141  0.6447141 0  0.5447141
#> 8  08 -0.4618396 -0.4618396 -0.3618396 1 -0.3618396
#> 9  09  0.6883405  0.6883405  0.7883405 1  0.7883405
#> 10 10 -0.2808157 -0.2808157 -0.1808157 1 -0.1808157

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.4315524   0.4315524   0.5315524  0.73155238  1.33155238  1  0
#> 2  02  0.1511179   0.1511179   0.2511179  0.45111791  1.05111791  0  0
#> 3  03 -0.2109492  -0.2109492  -0.1109492  0.08905078  0.68905078  1  1
#> 4  04  0.2996348   0.2996348   0.3996348  0.59963475  1.19963475  1  1
#> 5  05 -0.9736774  -0.9736774  -0.8736774 -0.67367735 -0.07367735  0  1
#> 6  06  0.6002301   0.6002301   0.7002301  0.90023007  1.50023007  1  0
#> 7  07 -1.4088598  -1.4088598  -1.3088598 -1.10885978 -0.50885978  1  0
#> 8  08  0.2252140   0.2252140   0.3252140  0.52521402  1.12521402  0  0
#> 9  09 -1.1838693  -1.1838693  -1.0838693 -0.88386931 -0.28386931  1  0
#> 10 10 -0.5282929  -0.5282929  -0.4282929 -0.22829294  0.37170706  1  1
#>             Y
#> 1   0.5315524
#> 2   0.1511179
#> 3   0.6890508
#> 4   1.1996348
#> 5  -0.6736774
#> 6   0.7002301
#> 7  -1.3088598
#> 8   0.2252140
#> 9  -1.0838693
#> 10  0.3717071