Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.16800261 -0.16800261 -0.06800261 1 -0.06800261
#> 2  02  0.03791973  0.03791973  0.13791973 0  0.03791973
#> 3  03  0.27137402  0.27137402  0.37137402 1  0.37137402
#> 4  04  1.31741721  1.31741721  1.41741721 1  1.41741721
#> 5  05  1.38962927  1.38962927  1.48962927 1  1.48962927
#> 6  06  0.58724470  0.58724470  0.68724470 0  0.58724470
#> 7  07 -1.08888643 -1.08888643 -0.98888643 1 -0.98888643
#> 8  08  0.03905162  0.03905162  0.13905162 0  0.03905162
#> 9  09  2.10408822  2.10408822  2.20408822 1  2.20408822
#> 10 10  0.36221498  0.36221498  0.46221498 1  0.46221498

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -1.505391566 -1.505391566 -1.40539157  -1.2053916  -0.6053916  1  1
#> 2  02 -2.519700789 -2.519700789 -2.41970079  -2.2197008  -1.6197008  1  0
#> 3  03 -0.177200590 -0.177200590 -0.07720059   0.1227994   0.7227994  0  1
#> 4  04  0.370614017  0.370614017  0.47061402   0.6706140   1.2706140  1  0
#> 5  05 -0.021455580 -0.021455580  0.07854442   0.2785444   0.8785444  0  1
#> 6  06 -1.553093819 -1.553093819 -1.45309382  -1.2530938  -0.6530938  0  1
#> 7  07  0.031055275  0.031055275  0.13105528   0.3310553   0.9310553  0  0
#> 8  08 -0.731493637 -0.731493637 -0.63149364  -0.4314936   0.1685064  0  1
#> 9  09 -1.091231334 -1.091231334 -0.99123133  -0.7912313  -0.1912313  0  1
#> 10 10  0.004902953  0.004902953  0.10490295   0.3049030   0.9049030  1  1
#>              Y
#> 1  -0.60539157
#> 2  -2.41970079
#> 3   0.12279941
#> 4   0.47061402
#> 5   0.27854442
#> 6  -1.25309382
#> 7   0.03105528
#> 8  -0.43149364
#> 9  -0.79123133
#> 10  0.90490295