Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0        Y_Z_1 Z           Y
#> 1  01  0.29557221  0.29557221  0.395572213 0  0.29557221
#> 2  02 -1.99733282 -1.99733282 -1.897332816 0 -1.99733282
#> 3  03 -0.29614315 -0.29614315 -0.196143155 0 -0.29614315
#> 4  04 -0.09534114 -0.09534114  0.004658859 0 -0.09534114
#> 5  05 -0.26612189 -0.26612189 -0.166121892 1 -0.16612189
#> 6  06 -0.55034252 -0.55034252 -0.450342523 1 -0.45034252
#> 7  07 -0.68201363 -0.68201363 -0.582013628 1 -0.58201363
#> 8  08  0.37849023  0.37849023  0.478490230 1  0.47849023
#> 9  09  0.20109035  0.20109035  0.301090354 0  0.20109035
#> 10 10 -0.88951478 -0.88951478 -0.789514776 0 -0.88951478

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.11686635  0.11686635  0.21686635   0.4168664   1.0168664  1  0
#> 2  02  0.95729005  0.95729005  1.05729005   1.2572901   1.8572901  1  0
#> 3  03 -0.44780010 -0.44780010 -0.34780010  -0.1478001   0.4521999  1  1
#> 4  04  0.70977577  0.70977577  0.80977577   1.0097758   1.6097758  0  0
#> 5  05  0.82381496  0.82381496  0.92381496   1.1238150   1.7238150  0  1
#> 6  06 -0.76728166 -0.76728166 -0.66728166  -0.4672817   0.1327183  0  0
#> 7  07 -0.01813469 -0.01813469  0.08186531   0.2818653   0.8818653  0  0
#> 8  08 -1.80951787 -1.80951787 -1.70951787  -1.5095179  -0.9095179  0  0
#> 9  09 -0.60149496 -0.60149496 -0.50149496  -0.3014950   0.2985050  0  0
#> 10 10 -0.31661710 -0.31661710 -0.21661710  -0.0166171   0.5833829  1  1
#>              Y
#> 1   0.21686635
#> 2   1.05729005
#> 3   0.45219990
#> 4   0.70977577
#> 5   1.12381496
#> 6  -0.76728166
#> 7  -0.01813469
#> 8  -1.80951787
#> 9  -0.60149496
#> 10  0.58338290