Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01  0.8739162  0.8739162  0.9739162 0  0.8739162
#> 2  02 -1.3628749 -1.3628749 -1.2628749 0 -1.3628749
#> 3  03 -0.5629859 -0.5629859 -0.4629859 0 -0.5629859
#> 4  04 -0.6606387 -0.6606387 -0.5606387 0 -0.6606387
#> 5  05  0.3642372  0.3642372  0.4642372 0  0.3642372
#> 6  06 -0.3039461 -0.3039461 -0.2039461 0 -0.3039461
#> 7  07  1.0461512  1.0461512  1.1461512 0  1.0461512
#> 8  08  0.8779778  0.8779778  0.9779778 0  0.8779778
#> 9  09  0.0839195  0.0839195  0.1839195 1  0.1839195
#> 10 10 -0.3150872 -0.3150872 -0.2150872 0 -0.3150872

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.51606018  0.51606018  0.61606018   0.8160602   1.4160602  0  0
#> 2  02 -0.07404308 -0.07404308  0.02595692   0.2259569   0.8259569  1  1
#> 3  03 -0.02119685 -0.02119685  0.07880315   0.2788031   0.8788031  0  1
#> 4  04  0.45120122  0.45120122  0.55120122   0.7512012   1.3512012  1  0
#> 5  05  1.60413914  1.60413914  1.70413914   1.9041391   2.5041391  1  1
#> 6  06  0.65340983  0.65340983  0.75340983   0.9534098   1.5534098  0  0
#> 7  07  2.02256974  2.02256974  2.12256974   2.3225697   2.9225697  1  1
#> 8  08 -0.75203749 -0.75203749 -0.65203749  -0.4520375   0.1479625  1  1
#> 9  09  0.36696213  0.36696213  0.46696213   0.6669621   1.2669621  1  1
#> 10 10 -0.45248742 -0.45248742 -0.35248742  -0.1524874   0.4475126  0  0
#>             Y
#> 1   0.5160602
#> 2   0.8259569
#> 3   0.2788031
#> 4   0.5512012
#> 5   2.5041391
#> 6   0.6534098
#> 7   2.9225697
#> 8   0.1479625
#> 9   1.2669621
#> 10 -0.4524874