Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -2.5280699 -2.5280699 -2.4280699 0 -2.5280699
#> 2  02  1.5319696  1.5319696  1.6319696 1  1.6319696
#> 3  03 -0.8312867 -0.8312867 -0.7312867 0 -0.8312867
#> 4  04 -1.0925710 -1.0925710 -0.9925710 0 -1.0925710
#> 5  05  0.8020168  0.8020168  0.9020168 0  0.8020168
#> 6  06  0.5402676  0.5402676  0.6402676 0  0.5402676
#> 7  07  0.6718521  0.6718521  0.7718521 1  0.7718521
#> 8  08  0.1687210  0.1687210  0.2687210 1  0.2687210
#> 9  09  0.8171137  0.8171137  0.9171137 0  0.8171137
#> 10 10 -1.5548087 -1.5548087 -1.4548087 0 -1.5548087

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.478721124  0.478721124  0.57872112   0.7787211  1.37872112  1  0
#> 2  02 -1.868822558 -1.868822558 -1.76882256  -1.5688226 -0.96882256  0  1
#> 3  03 -0.755813692 -0.755813692 -0.65581369  -0.4558137  0.14418631  1  0
#> 4  04 -0.570150120 -0.570150120 -0.47015012  -0.2701501  0.32984988  1  1
#> 5  05 -1.812901909 -1.812901909 -1.71290191  -1.5129019 -0.91290191  1  1
#> 6  06 -0.001043247 -0.001043247  0.09895675   0.2989568  0.89895675  0  0
#> 7  07 -0.745026270 -0.745026270 -0.64502627  -0.4450263  0.15497373  0  1
#> 8  08 -1.576697911 -1.576697911 -1.47669791  -1.2766979 -0.67669791  0  1
#> 9  09 -0.928162491 -0.928162491 -0.82816249  -0.6281625 -0.02816249  1  0
#> 10 10  1.354439919  1.354439919  1.45443992   1.6544399  2.25443992  0  1
#>               Y
#> 1   0.578721124
#> 2  -1.568822558
#> 3  -0.655813692
#> 4   0.329849880
#> 5  -0.912901909
#> 6  -0.001043247
#> 7  -0.445026270
#> 8  -1.276697911
#> 9  -0.828162491
#> 10  1.654439919