Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -0.9376118 -0.9376118 -0.8376118 0 -0.9376118
#> 2  02  1.7498372  1.7498372  1.8498372 1  1.8498372
#> 3  03  0.2022991  0.2022991  0.3022991 1  0.3022991
#> 4  04 -0.4758486 -0.4758486 -0.3758486 1 -0.3758486
#> 5  05 -1.7597746 -1.7597746 -1.6597746 0 -1.7597746
#> 6  06 -0.8475088 -0.8475088 -0.7475088 0 -0.8475088
#> 7  07  1.3972215  1.3972215  1.4972215 0  1.3972215
#> 8  08 -0.7591654 -0.7591654 -0.6591654 1 -0.6591654
#> 9  09  1.4855211  1.4855211  1.5855211 0  1.4855211
#> 10 10 -0.8930730 -0.8930730 -0.7930730 1 -0.7930730

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.36787382  0.36787382  0.46787382   0.6678738   1.2678738  1  1
#> 2  02 -0.43211972 -0.43211972 -0.33211972  -0.1321197   0.4678803  1  0
#> 3  03  0.61333721  0.61333721  0.71333721   0.9133372   1.5133372  0  0
#> 4  04  1.44054707  1.44054707  1.54054707   1.7405471   2.3405471  0  0
#> 5  05  0.84249275  0.84249275  0.94249275   1.1424928   1.7424928  1  0
#> 6  06 -1.74633844 -1.74633844 -1.64633844  -1.4463384  -0.8463384  0  1
#> 7  07 -0.17117574 -0.17117574 -0.07117574   0.1288243   0.7288243  0  0
#> 8  08  0.40429588  0.40429588  0.50429588   0.7042959   1.3042959  0  1
#> 9  09  0.32483347  0.32483347  0.42483347   0.6248335   1.2248335  0  1
#> 10 10  0.04931281  0.04931281  0.14931281   0.3493128   0.9493128  0  1
#>             Y
#> 1   1.2678738
#> 2  -0.3321197
#> 3   0.6133372
#> 4   1.4405471
#> 5   0.9424928
#> 6  -1.4463384
#> 7  -0.1711757
#> 8   0.7042959
#> 9   0.6248335
#> 10  0.3493128