Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z          Y
#> 1  01  1.3972036  1.3972036  1.49720361 1  1.4972036
#> 2  02  0.6845201  0.6845201  0.78452007 1  0.7845201
#> 3  03  0.7420112  0.7420112  0.84201117 0  0.7420112
#> 4  04  2.3650433  2.3650433  2.46504332 0  2.3650433
#> 5  05  0.8627795  0.8627795  0.96277949 0  0.8627795
#> 6  06  0.3223356  0.3223356  0.42233557 1  0.4223356
#> 7  07 -1.4864857 -1.4864857 -1.38648573 0 -1.4864857
#> 8  08 -0.1197522 -0.1197522 -0.01975218 0 -0.1197522
#> 9  09  0.7631570  0.7631570  0.86315702 1  0.8631570
#> 10 10  0.7670986  0.7670986  0.86709864 1  0.8670986

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0   Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.2603578   1.2603578  1.36035775  1.5603577513   2.1603578  1  1
#> 2  02 -1.7352201  -1.7352201 -1.63522011 -1.4352201091  -0.8352201  0  0
#> 3  03  1.4697031   1.4697031  1.56970306  1.7697030624   2.3697031  1  1
#> 4  04 -0.1403856  -0.1403856 -0.04038559  0.1596144075   0.7596144  1  1
#> 5  05 -0.1835938  -0.1835938 -0.08359378  0.1164062200   0.7164062  0  0
#> 6  06  1.7548720   1.7548720  1.85487204  2.0548720390   2.6548720  1  0
#> 7  07 -0.2997641  -0.2997641 -0.19976408  0.0002359162   0.6002359  0  0
#> 8  08 -1.0599245  -1.0599245 -0.95992451 -0.7599245075  -0.1599245  1  0
#> 9  09 -0.3848464  -0.3848464 -0.28484637 -0.0848463691   0.5151536  0  0
#> 10 10  2.0909653   2.0909653  2.19096530  2.3909652999   2.9909653  0  1
#>             Y
#> 1   2.1603578
#> 2  -1.7352201
#> 3   2.3697031
#> 4   0.7596144
#> 5  -0.1835938
#> 6   1.8548720
#> 7  -0.2997641
#> 8  -0.9599245
#> 9  -0.3848464
#> 10  2.3909653