Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z          Y
#> 1  01 -0.87558804 -0.87558804 -0.7755880 1 -0.7755880
#> 2  02  1.92946735  1.92946735  2.0294674 0  1.9294674
#> 3  03  0.14751214  0.14751214  0.2475121 0  0.1475121
#> 4  04 -1.00148876 -1.00148876 -0.9014888 1 -0.9014888
#> 5  05  0.04450922  0.04450922  0.1445092 1  0.1445092
#> 6  06 -1.05156362 -1.05156362 -0.9515636 0 -1.0515636
#> 7  07 -0.35518545 -0.35518545 -0.2551854 0 -0.3551854
#> 8  08 -1.42391190 -1.42391190 -1.3239119 1 -1.3239119
#> 9  09  0.83352477  0.83352477  0.9335248 1  0.9335248
#> 10 10 -0.29024519 -0.29024519 -0.1902452 1 -0.1902452

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.06786762  1.06786762  1.16786762   1.3678676  1.96786762  1  0
#> 2  02  0.42686900  0.42686900  0.52686900   0.7268690  1.32686900  1  0
#> 3  03 -0.77530305 -0.77530305 -0.67530305  -0.4753031  0.12469695  0  1
#> 4  04 -1.41059889 -1.41059889 -1.31059889  -1.1105989 -0.51059889  0  1
#> 5  05 -0.98095221 -0.98095221 -0.88095221  -0.6809522 -0.08095221  1  1
#> 6  06 -0.83730108 -0.83730108 -0.73730108  -0.5373011  0.06269892  0  0
#> 7  07 -1.23768017 -1.23768017 -1.13768017  -0.9376802 -0.33768017  0  0
#> 8  08 -0.94619014 -0.94619014 -0.84619014  -0.6461901 -0.04619014  1  1
#> 9  09 -0.07490869 -0.07490869  0.02509131   0.2250913  0.82509131  0  1
#> 10 10 -0.56852253 -0.56852253 -0.46852253  -0.2685225  0.33147747  1  0
#>              Y
#> 1   1.16786762
#> 2   0.52686900
#> 3  -0.47530305
#> 4  -1.11059889
#> 5  -0.08095221
#> 6  -0.83730108
#> 7  -1.23768017
#> 8  -0.04619014
#> 9   0.22509131
#> 10 -0.46852253