Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z           Y
#> 1  01  0.11200394  0.11200394  0.2120039 1  0.21200394
#> 2  02 -0.69692457 -0.69692457 -0.5969246 1 -0.59692457
#> 3  03  1.84210005  1.84210005  1.9421001 0  1.84210005
#> 4  04 -1.86560766 -1.86560766 -1.7656077 0 -1.86560766
#> 5  05 -0.26550280 -0.26550280 -0.1655028 1 -0.16550280
#> 6  06  0.04765361  0.04765361  0.1476536 0  0.04765361
#> 7  07 -0.40182573 -0.40182573 -0.3018257 0 -0.40182573
#> 8  08 -0.57101885 -0.57101885 -0.4710188 0 -0.57101885
#> 9  09  0.17820182  0.17820182  0.2782018 0  0.17820182
#> 10 10 -0.60079680 -0.60079680 -0.5007968 0 -0.60079680

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.1782552   1.1782552   1.2782552   1.4782552  2.07825516  0  1
#> 2  02  0.8894423   0.8894423   0.9894423   1.1894423  1.78944231  1  1
#> 3  03  0.1180383   0.1180383   0.2180383   0.4180383  1.01803832  1  0
#> 4  04 -0.9990234  -0.9990234  -0.8990234  -0.6990234 -0.09902345  0  0
#> 5  05 -1.2318663  -1.2318663  -1.1318663  -0.9318663 -0.33186634  0  0
#> 6  06  0.5539301   0.5539301   0.6539301   0.8539301  1.45393012  1  1
#> 7  07  1.0335082   1.0335082   1.1335082   1.3335082  1.93350819  1  1
#> 8  08 -1.3113323  -1.3113323  -1.2113323  -1.0113323 -0.41133233  1  1
#> 9  09  0.8117456   0.8117456   0.9117456   1.1117456  1.71174555  1  1
#> 10 10 -0.4691305  -0.4691305  -0.3691305  -0.1691305  0.43086947  1  0
#>             Y
#> 1   1.4782552
#> 2   1.7894423
#> 3   0.2180383
#> 4  -0.9990234
#> 5  -1.2318663
#> 6   1.4539301
#> 7   1.9335082
#> 8  -0.4113323
#> 9   1.7117456
#> 10 -0.3691305