Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID            U        Y_Z_0       Y_Z_1 Z           Y
#> 1  01 -0.117354410 -0.117354410 -0.01735441 1 -0.01735441
#> 2  02  1.770316388  1.770316388  1.87031639 1  1.87031639
#> 3  03 -0.940821920 -0.940821920 -0.84082192 1 -0.84082192
#> 4  04 -2.207414933 -2.207414933 -2.10741493 1 -2.10741493
#> 5  05  0.541472952  0.541472952  0.64147295 0  0.54147295
#> 6  06  2.254748593  2.254748593  2.35474859 0  2.25474859
#> 7  07  0.002202776  0.002202776  0.10220278 1  0.10220278
#> 8  08 -0.201087145 -0.201087145 -0.10108714 0 -0.20108714
#> 9  09 -0.587752190 -0.587752190 -0.48775219 1 -0.48775219
#> 10 10  1.387012708  1.387012708  1.48701271 0  1.38701271

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  0.06642001  0.06642001  0.16642001   0.3664200  0.96642001  0  0
#> 2  02  0.14407404  0.14407404  0.24407404   0.4440740  1.04407404  1  0
#> 3  03 -0.19964164 -0.19964164 -0.09964164   0.1003584  0.70035836  0  0
#> 4  04  1.06862819  1.06862819  1.16862819   1.3686282  1.96862819  1  0
#> 5  05 -1.27390313 -1.27390313 -1.17390313  -0.9739031 -0.37390313  1  0
#> 6  06 -0.81385408 -0.81385408 -0.71385408  -0.5138541  0.08614592  0  1
#> 7  07 -1.35971924 -1.35971924 -1.25971924  -1.0597192 -0.45971924  1  0
#> 8  08 -0.80269544 -0.80269544 -0.70269544  -0.5026954  0.09730456  0  0
#> 9  09 -1.18071298 -1.18071298 -1.08071298  -0.8807130 -0.28071298  0  1
#> 10 10 -2.34381647 -2.34381647 -2.24381647  -2.0438165 -1.44381647  0  0
#>              Y
#> 1   0.06642001
#> 2   0.24407404
#> 3  -0.19964164
#> 4   1.16862819
#> 5  -1.17390313
#> 6  -0.51385408
#> 7  -1.25971924
#> 8  -0.80269544
#> 9  -0.88071298
#> 10 -2.34381647