Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0       Y_Z_1 Z           Y
#> 1  01  2.2647222  2.2647222  2.36472225 0  2.26472225
#> 2  02 -1.5201457 -1.5201457 -1.42014569 1 -1.42014569
#> 3  03 -0.3861454 -0.3861454 -0.28614543 1 -0.28614543
#> 4  04  0.3971820  0.3971820  0.49718202 0  0.39718202
#> 5  05 -0.1829165 -0.1829165 -0.08291651 1 -0.08291651
#> 6  06  0.9985643  0.9985643  1.09856431 0  0.99856431
#> 7  07 -0.1388638 -0.1388638 -0.03886375 1 -0.03886375
#> 8  08 -0.5830693 -0.5830693 -0.48306928 1 -0.48306928
#> 9  09 -1.6875512 -1.6875512 -1.58755120 1 -1.58755120
#> 10 10  0.3254820  0.3254820  0.42548204 0  0.32548204

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID            U  Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.174390703 -0.174390703 -0.07439070   0.1256093   0.7256093  1  0
#> 2  02  0.857613028  0.857613028  0.95761303   1.1576130   1.7576130  0  1
#> 3  03  0.580275947  0.580275947  0.68027595   0.8802759   1.4802759  1  0
#> 4  04  0.861472678  0.861472678  0.96147268   1.1614727   1.7614727  0  1
#> 5  05 -0.713825808 -0.713825808 -0.61382581  -0.4138258   0.1861742  0  1
#> 6  06  0.005900351  0.005900351  0.10590035   0.3059004   0.9059004  1  0
#> 7  07  0.042801739  0.042801739  0.14280174   0.3428017   0.9428017  1  0
#> 8  08  0.235652114  0.235652114  0.33565211   0.5356521   1.1356521  1  0
#> 9  09 -0.009386273 -0.009386273  0.09061373   0.2906137   0.8906137  0  1
#> 10 10  1.059439812  1.059439812  1.15943981   1.3594398   1.9594398  1  1
#>             Y
#> 1  -0.0743907
#> 2   1.1576130
#> 3   0.6802759
#> 4   1.1614727
#> 5  -0.4138258
#> 6   0.1059004
#> 7   0.1428017
#> 8   0.3356521
#> 9   0.2906137
#> 10  1.9594398