Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  0.52140819  0.52140819  0.62140819 1  0.62140819
#> 2  02 -0.14530583 -0.14530583 -0.04530583 0 -0.14530583
#> 3  03  0.65986873  0.65986873  0.75986873 0  0.65986873
#> 4  04 -0.79658992 -0.79658992 -0.69658992 1 -0.69658992
#> 5  05  1.61774944  1.61774944  1.71774944 1  1.71774944
#> 6  06  0.40128173  0.40128173  0.50128173 0  0.40128173
#> 7  07 -0.78774856 -0.78774856 -0.68774856 0 -0.78774856
#> 8  08 -0.25355587 -0.25355587 -0.15355587 1 -0.15355587
#> 9  09 -0.02492597 -0.02492597  0.07507403 0 -0.02492597
#> 10 10 -1.35431958 -1.35431958 -1.25431958 0 -1.35431958

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.9890505   1.9890505  2.08905049  2.28905049   2.8890505  0  0
#> 2  02 -0.4999128  -0.4999128 -0.39991282 -0.19991282   0.4000872  1  0
#> 3  03 -0.1409879  -0.1409879 -0.04098788  0.15901212   0.7590121  1  0
#> 4  04  0.5004166   0.5004166  0.60041662  0.80041662   1.4004166  0  1
#> 5  05 -1.2310227  -1.2310227 -1.13102270 -0.93102270  -0.3310227  1  1
#> 6  06 -0.2813094  -0.2813094 -0.18130945  0.01869055   0.6186906  1  1
#> 7  07  1.4670970   1.4670970  1.56709700  1.76709700   2.3670970  1  1
#> 8  08 -0.1517265  -0.1517265 -0.05172654  0.14827346   0.7482735  1  0
#> 9  09 -0.2875687  -0.2875687 -0.18756870  0.01243130   0.6124313  0  0
#> 10 10  1.1212063   1.1212063  1.22120625  1.42120625   2.0212063  1  0
#>              Y
#> 1   1.98905049
#> 2  -0.39991282
#> 3  -0.04098788
#> 4   0.80041662
#> 5  -0.33102270
#> 6   0.61869055
#> 7   2.36709700
#> 8  -0.05172654
#> 9  -0.28756870
#> 10  1.22120625