Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0      Y_Z_1 Z           Y
#> 1  01  0.84057062  0.84057062  0.9405706 1  0.94057062
#> 2  02  0.06234736  0.06234736  0.1623474 0  0.06234736
#> 3  03  1.25021879  1.25021879  1.3502188 0  1.25021879
#> 4  04  0.54577328  0.54577328  0.6457733 1  0.64577328
#> 5  05  0.62665296  0.62665296  0.7266530 0  0.62665296
#> 6  06 -0.61081445 -0.61081445 -0.5108144 0 -0.61081445
#> 7  07  0.31848086  0.31848086  0.4184809 0  0.31848086
#> 8  08 -0.64928498 -0.64928498 -0.5492850 1 -0.54928498
#> 9  09  0.37876949  0.37876949  0.4787695 0  0.37876949
#> 10 10  0.73523232  0.73523232  0.8352323 1  0.83523232

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID           U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.77655646 -0.77655646 -0.67655646  -0.4765565  0.12344354  0  1
#> 2  02  0.24206339  0.24206339  0.34206339   0.5420634  1.14206339  0  1
#> 3  03 -0.87628491 -0.87628491 -0.77628491  -0.5762849  0.02371509  1  0
#> 4  04 -1.01979473 -1.01979473 -0.91979473  -0.7197947 -0.11979473  1  0
#> 5  05  0.11155511  0.11155511  0.21155511   0.4115551  1.01155511  1  0
#> 6  06  1.50127817  1.50127817  1.60127817   1.8012782  2.40127817  0  1
#> 7  07 -0.02194891 -0.02194891  0.07805109   0.2780511  0.87805109  0  0
#> 8  08 -0.02795654 -0.02795654  0.07204346   0.2720435  0.87204346  1  1
#> 9  09 -2.28348293 -2.28348293 -2.18348293  -1.9834829 -1.38348293  0  1
#> 10 10 -0.53305651 -0.53305651 -0.43305651  -0.2330565  0.36694349  1  0
#>              Y
#> 1  -0.47655646
#> 2   0.54206339
#> 3  -0.77628491
#> 4  -0.91979473
#> 5   0.21155511
#> 6   1.80127817
#> 7  -0.02194891
#> 8   0.87204346
#> 9  -1.98348293
#> 10 -0.43305651