Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID          U      Y_Z_0      Y_Z_1 Z          Y
#> 1  01  0.3869117  0.3869117  0.4869117 0  0.3869117
#> 2  02 -0.2870729 -0.2870729 -0.1870729 1 -0.1870729
#> 3  03  0.2736687  0.2736687  0.3736687 0  0.2736687
#> 4  04  1.0243016  1.0243016  1.1243016 1  1.1243016
#> 5  05 -0.4755767 -0.4755767 -0.3755767 1 -0.3755767
#> 6  06  1.8175314  1.8175314  1.9175314 1  1.9175314
#> 7  07 -2.0141480 -2.0141480 -1.9141480 1 -1.9141480
#> 8  08  0.1045993  0.1045993  0.2045993 0  0.1045993
#> 9  09 -0.6095905 -0.6095905 -0.5095905 0 -0.6095905
#> 10 10  0.6620271  0.6620271  0.7620271 0  0.6620271

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01 -0.5944686  -0.5944686 -0.49446858  -0.2944686   0.3055314  1  0
#> 2  02 -0.1628354  -0.1628354 -0.06283536   0.1371646   0.7371646  0  1
#> 3  03  0.5728740   0.5728740  0.67287400   0.8728740   1.4728740  0  0
#> 4  04  1.1354428   1.1354428  1.23544281   1.4354428   2.0354428  1  1
#> 5  05  0.1011208   0.1011208  0.20112085   0.4011208   1.0011208  0  0
#> 6  06 -1.0714457  -1.0714457 -0.97144570  -0.7714457  -0.1714457  1  1
#> 7  07 -1.7401791  -1.7401791 -1.64017907  -1.4401791  -0.8401791  1  1
#> 8  08  0.2888387   0.2888387  0.38883868   0.5888387   1.1888387  0  0
#> 9  09 -0.1638437  -0.1638437 -0.06384371   0.1361563   0.7361563  0  0
#> 10 10  0.2680544   0.2680544  0.36805438   0.5680544   1.1680544  0  0
#>             Y
#> 1  -0.4944686
#> 2   0.1371646
#> 3   0.5728740
#> 4   2.0354428
#> 5   0.1011208
#> 6  -0.1714457
#> 7  -0.8401791
#> 8   0.2888387
#> 9  -0.1638437
#> 10  0.2680544