Reveal outcomes

Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.

Usage

reveal_outcomes(x)

Arguments

x

A formula with the outcome name on the left hand side and assignment variables on the right hand side (e.g., Y ~ Z).

Examples

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

fabricate(
  data = dat,
  Z = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z)
)
#>    ID           U       Y_Z_0       Y_Z_1 Z           Y
#> 1  01  1.20222273  1.20222273  1.30222273 1  1.30222273
#> 2  02 -1.25249675 -1.25249675 -1.15249675 1 -1.15249675
#> 3  03 -0.56865110 -0.56865110 -0.46865110 0 -0.56865110
#> 4  04 -0.08447523 -0.08447523  0.01552477 0 -0.08447523
#> 5  05  0.29724592  0.29724592  0.39724592 1  0.39724592
#> 6  06 -0.58284920 -0.58284920 -0.48284920 0 -0.58284920
#> 7  07  0.39076162  0.39076162  0.49076162 1  0.49076162
#> 8  08  1.37880358  1.37880358  1.47880358 0  1.37880358
#> 9  09  0.92087699  0.92087699  1.02087699 1  1.02087699
#> 10 10 -0.14267835 -0.14267835 -0.04267835 1 -0.04267835

fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z1 + 0.3 * Z2 + 0.5 * Z1 * Z2 + U,
                     conditions = list(Z1 = c(0, 1),
                                       Z2 = c(0, 1))),
  Z1 = rbinom(N, 1, prob = 0.5),
  Z2 = rbinom(N, 1, prob = 0.5),
  Y = reveal_outcomes(Y ~ Z1 + Z2)
)
#>    ID          U Y_Z1_0_Z2_0 Y_Z1_1_Z2_0 Y_Z1_0_Z2_1 Y_Z1_1_Z2_1 Z1 Z2
#> 1  01  1.1815185   1.1815185  1.28151852   1.4815185   2.0815185  0  1
#> 2  02  1.0763269   1.0763269  1.17632688   1.3763269   1.9763269  1  0
#> 3  03  0.2088421   0.2088421  0.30884211   0.5088421   1.1088421  0  1
#> 4  04  1.5015288   1.5015288  1.60152882   1.8015288   2.4015288  0  0
#> 5  05 -1.0467563  -1.0467563 -0.94675627  -0.7467563  -0.1467563  0  0
#> 6  06 -1.0668400  -1.0668400 -0.96683997  -0.7668400  -0.1668400  1  1
#> 7  07  2.2637315   2.2637315  2.36373154   2.5637315   3.1637315  1  0
#> 8  08 -1.1066274  -1.1066274 -1.00662736  -0.8066274  -0.2066274  0  0
#> 9  09  0.3596287   0.3596287  0.45962875   0.6596287   1.2596287  0  0
#> 10 10 -0.1347115  -0.1347115 -0.03471148   0.1652885   0.7652885  1  1
#>             Y
#> 1   1.4815185
#> 2   1.1763269
#> 3   0.5088421
#> 4   1.5015288
#> 5  -1.0467563
#> 6  -0.1668400
#> 7   2.3637315
#> 8  -1.1066274
#> 9   0.3596287
#> 10  0.7652885